Parkinsons data from UCI machine learning archive

Biomedical voice measurements from 31 people, 23 with Parkinson's disease (PD). Each column in the table is a particular voice measure, and each row corresponds one of 195 voice recording from these individuals ("name" column). The main aim of the data is to discriminate healthy people from those with PD, according to "status" column which is set to 0 for healthy and 1 for PD.

Format

A tibble with 1,013 observations and 3 variables

name

ASCII subject name and recording number

MDVP:Fo(Hz)

Average vocal fundamental frequency

MDVP:Fhi(Hz)

Maximum vocal fundamental frequency

MDVP:Flo(Hz)

Minimum vocal fundamental frequency

MDVP:Jitter,MDVP:Jitter(Abs),MDVP:RAP,MDVP:PPQ,Jitter:DDP

Several measures of variation in fundamental frequency

MDVP:Shimmer,MDVP:Shimmer(dB),Shimmer:APQ3,Shimmer:APQ5,MDVP:APQ,Shimmer:DDA

Several measures of variation in amplitude

NHR,HNR

Two measures of ratio of noise to tonal components in the voice

status

Health status of the subject (one) - Parkinson's, (zero) - healthy

RPDE,D2

Two nonlinear dynamical complexity measures

DFA

Signal fractal scaling exponent

spread1,spread2,PPE

Three nonlinear measures of fundamental frequency variation

Details

The data is available at The UCI Machine Learning Repository in ASCII CSV format. The rows of the CSV file contain an instance corresponding to one voice recording. There are around six recordings per patient, the name of the patient is identified in the first column.

The data are originally analysed in: Max A. Little, Patrick E. McSharry, Eric J. Hunter, Lorraine O. Ramig (2008), 'Suitability of dysphonia measurements for telemonitoring of Parkinson's disease', IEEE Transactions on Biomedical Engineering.