Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1           y1           x2           y2          mx1
#>    62   79 0.3634317382 0.5959901301 0.3444601990 0.6052700307  0.338926803
#>     1   79 0.3054776388 0.5572361150 0.3444601990 0.6052700307  0.274846915
#>    29    7 0.2650356960 0.4624196007 0.4636868728 0.4278440143  0.369376830
#>    36   29 0.1936345378 0.5287609463 0.2650356960 0.4624196007  0.254680986
#>    66   36 0.2163473860 0.5806824935 0.1936345378 0.5287609463  0.251782726
#>    23   12 0.4230200637 0.7034241702 0.3761285772 0.7353842626  0.365295040
#>    12   77 0.3761285772 0.7353842626 0.3371242858 0.7768486278  0.298433353
#>    62    7 0.3634317382 0.5959901301 0.4636868728 0.4278440143  0.449797151
#>    28    7 0.4675944762 0.3370529010 0.4636868728 0.4278440143  0.352604371
#>    85   23 0.5002122045 0.6716518611 0.4230200637 0.7034241702  0.436227842
#>    13   15 0.5131683391 0.7104583822 0.5030238466 0.7325563598  0.469099638
#>     2   85 0.5162246064 0.6365204073 0.5002122045 0.6716518611  0.438252171
#>    79   12 0.3444601990 0.6052700307 0.3761285772 0.7353842626  0.296376211
#>    23   75 0.4230200637 0.7034241702 0.4935387162 0.8213753645  0.437390875
#>    19   90 0.6353272900 0.5302164045 0.6492330841 0.5583215095  0.598312304
#>     2   62 0.5162246064 0.6365204073 0.3634317382 0.5959901301  0.453318769
#>    16   36 0.0911648673 0.5134230284 0.1936345378 0.5287609463  0.132641050
#>    66   79 0.2163473860 0.5806824935 0.3444601990 0.6052700307  0.266834402
#>    15   23 0.5030238466 0.7325563598 0.4230200637 0.7034241702  0.468245666
#>    13   85 0.5131683391 0.7104583822 0.5002122045 0.6716518611  0.554389224
#>    80   36 0.2037706969 0.4524766260 0.1936345378 0.5287609463  0.147970020
#>    28   29 0.4675944762 0.3370529010 0.2650356960 0.4624196007  0.348851188
#>    43   26 0.2323059930 0.7544823097 0.2141102154 0.8078739020  0.155968412
#>    43   77 0.2323059930 0.7544823097 0.3371242858 0.7768486278  0.277459828
#>    71   86 0.0855045795 0.7286560312 0.0491712449 0.5941819538  0.137209309
#>    16   86 0.0911648673 0.5134230284 0.0491712449 0.5941819538  0.113353797
#>    77   75 0.3371242858 0.7768486278 0.4935387162 0.8213753645  0.411558375
#>    90    2 0.6492330841 0.5583215095 0.5162246064 0.6365204073  0.612216741
#>    43   71 0.2323059930 0.7544823097 0.0855045795 0.7286560312  0.171061082
#>    90    5 0.6492330841 0.5583215095 0.5713509127 0.7741688117  0.626665694
#>     7   21 0.4636868728 0.4278440143 0.6092004780 0.3947836554  0.551120271
#>    16   94 0.0911648673 0.5134230284 0.0434441415 0.5009548594  0.086661030
#>    80   16 0.2037706969 0.4524766260 0.0911648673 0.5134230284  0.121541127
#>    16   89 0.0911648673 0.5134230284 0.0475215113 0.5659260962  0.060695100
#>    86   18 0.0491712449 0.5941819538 0.0222226822 0.6822887040  0.034584755
#>    75   15 0.4935387162 0.8213753645 0.5030238466 0.7325563598  0.521314508
#>    41   33 0.3287631143 0.8800476613 0.1059619109 0.8614958192  0.211692327
#>    41   26 0.3287631143 0.8800476613 0.2141102154 0.8078739020  0.283712708
#>     5   13 0.5713509127 0.7741688117 0.5131683391 0.7104583822  0.613165738
#>    29   80 0.2650356960 0.4624196007 0.2037706969 0.4524766260  0.251452271
#>    21   39 0.6092004780 0.3947836554 0.7013501159 0.3908462874  0.657958574
#>    21   19 0.6092004780 0.3947836554 0.6353272900 0.5302164045  0.551706546
#>    28   22 0.4675944762 0.3370529010 0.4336725378 0.2467560993  0.469893456
#>    81   80 0.3056233310 0.2556777984 0.2037706969 0.4524766260  0.217108159
#>    77   78 0.3371242858 0.7768486278 0.4608184374 0.8785130097  0.397386410
#>    75   24 0.4935387162 0.8213753645 0.5866144017 0.8216949868  0.540065821
#>    89   86 0.0475215113 0.5659260962 0.0491712449 0.5941819538  0.020378088
#>    81   22 0.3056233310 0.2556777984 0.4336725378 0.2467560993  0.374476181
#>    94   89 0.0434441415 0.5009548594 0.0475215113 0.5659260962 -0.003516782
#>    89   47 0.0475215113 0.5659260962 0.0009797839 0.6063623888 -0.013973419
#>    41   77 0.3287631143 0.8800476613 0.3371242858 0.7768486278  0.388136203
#>    94   84 0.0434441415 0.5009548594 0.0000000000 0.4777148316  0.057305003
#>     5   24 0.5713509127 0.7741688117 0.5866144017 0.8216949868  0.540114747
#>    90  100 0.6492330841 0.5583215095 0.7259117501 0.5229049673  0.694885034
#>    18   71 0.0222226822 0.6822887040 0.0855045795 0.7286560312  0.037946749
#>    26   33 0.2141102154 0.8078739020 0.1059619109 0.8614958192  0.137228789
#>    33   25 0.1059619109 0.8614958192 0.0498017708 0.7732753444  0.118414387
#>    21   28 0.6092004780 0.3947836554 0.4675944762 0.3370529010  0.549247320
#>    93   91 0.5351184728 0.8710421744 0.4872104806 0.9042115395  0.497237913
#>    78   75 0.4608184374 0.8785130097 0.4935387162 0.8213753645  0.496589135
#>    91   54 0.4872104806 0.9042115395 0.3842868552 0.9086257009  0.436744496
#>    81   37 0.3056233310 0.2556777984 0.1000600514 0.3537818960  0.212031341
#>    32   38 0.0325758348 0.3474112355 0.1006092894 0.2921967258  0.068950552
#>    54   33 0.3842868552 0.9086257009 0.1059619109 0.8614958192  0.188163371
#>    37   94 0.1000600514 0.3537818960 0.0434441415 0.5009548594  0.059514913
#>    58   90 0.7769132417 0.7309053745 0.6492330841 0.5583215095  0.700273205
#>    59   34 0.5180089228 0.2638303876 0.5682827467 0.2375056718  0.554654935
#>    58    5 0.7769132417 0.7309053745 0.5713509127 0.7741688117  0.659724647
#>    21   59 0.6092004780 0.3947836554 0.5180089228 0.2638303876  0.571286122
#>    59   22 0.5180089228 0.2638303876 0.4336725378 0.2467560993  0.495282836
#>    39   50 0.7013501159 0.3908462874 0.8106550740 0.4585176575  0.738991520
#>    39  100 0.7013501159 0.3908462874 0.7259117501 0.5229049673  0.675566384
#>    45   27 0.5791605126 0.2440093515 0.7219315716 0.2561021829  0.645539538
#>    35   22 0.3226057961 0.1460559555 0.4336725378 0.2467560993  0.366707522
#>    51   35 0.2494884225 0.1623540158 0.3226057961 0.1460559555  0.295816299
#>    81   38 0.3056233310 0.2556777984 0.1006092894 0.2921967258  0.199875107
#>    38   61 0.1006092894 0.2921967258 0.0152284580 0.3034370101  0.053915661
#>    40   93 0.6629957170 0.8571073676 0.5351184728 0.8710421744  0.601780156
#>    99   40 0.6725360499 0.8211587877 0.6629957170 0.8571073676  0.629622550
#>    32   61 0.0325758348 0.3474112355 0.0152284580 0.3034370101  0.059694493
#>    32   84 0.0325758348 0.3474112355 0.0000000000 0.4777148316 -0.114257609
#>    57   34 0.3201244288 0.0930332358 0.5682827467 0.2375056718  0.490259318
#>    17   91 0.6264319693 0.9769434109 0.4872104806 0.9042115395  0.552952419
#>    51   52 0.2494884225 0.1623540158 0.1002362212 0.1711840128  0.176344928
#>    38   46 0.1006092894 0.2921967258 0.0013167037 0.2621080528  0.064706004
#>    61   84 0.0152284580 0.3034370101 0.0000000000 0.4777148316 -0.246121956
#>    38   52 0.1006092894 0.2921967258 0.1002362212 0.1711840128  0.178688897
#>    99    5 0.6725360499 0.8211587877 0.5713509127 0.7741688117  0.664391992
#>     6   83 0.8415481020 0.5709817379 0.8303568039 0.6036763120  0.788896344
#>    11   82 0.8057000284 0.4852282456 0.7402570598 0.5510603195  0.771988965
#>    50   11 0.8106550740 0.4585176575 0.8057000284 0.4852282456  0.745076157
#>    35   57 0.3226057961 0.1460559555 0.3201244288 0.0930332358  0.453430157
#>    35    8 0.3226057961 0.1460559555 0.2511959396 0.1423676212  0.288094514
#>    51   98 0.2494884225 0.1623540158 0.1171151995 0.1397658550  0.180553451
#>    90   82 0.6492330841 0.5583215095 0.7402570598 0.5510603195  0.702405926
#>     8   51 0.2511959396 0.1423676212 0.2494884225 0.1623540158  0.197215262
#>    42   98 0.0816309661 0.0478530301 0.1171151995 0.1397658550  0.037997847
#>    27   39 0.7219315716 0.2561021829 0.7013501159 0.3908462874  0.651922759
#>    58   31 0.7769132417 0.7309053745 0.8086489416 0.6595662310  0.768746259
#>    31   82 0.8086489416 0.6595662310 0.7402570598 0.5510603195  0.765804362
#>    83   14 0.8303568039 0.6036763120 0.8157441382 0.6872455277  0.881406504
#>    58   10 0.7769132417 0.7309053745 0.8111275515 0.7102307245  0.798942239
#>    58   14 0.7769132417 0.7309053745 0.8157441382 0.6872455277  0.776325217
#>    76   45 0.7099775598 0.2366303992 0.5791605126 0.2440093515  0.641316151
#>    40   17 0.6629957170 0.8571073676 0.6264319693 0.9769434109  0.603455354
#>    99   58 0.6725360499 0.8211587877 0.7769132417 0.7309053745  0.773012893
#>    55   97 0.7257551286 0.0813661337 0.7278531507 0.0914161440  0.600695759
#>    57    8 0.3201244288 0.0930332358 0.2511959396 0.1423676212  0.246425877
#>     9   74 0.9364079020 0.7640727483 0.8791066229 0.7605102239  0.910827831
#>     6   68 0.8415481020 0.5709817379 0.8583518018 0.5423971861  0.937408364
#>    11    6 0.8057000284 0.4852282456 0.8415481020 0.5709817379  0.795025006
#>    70   17 0.8165272772 0.9546310835 0.6264319693 0.9769434109  0.718394945
#>    14   74 0.8157441382 0.6872455277 0.8791066229 0.7605102239  0.864609432
#>     6   87 0.8415481020 0.5709817379 1.0000000000 0.6896468422  0.929090158
#>    27   95 0.7219315716 0.2561021829 0.7272713816 0.1981375014  0.826473244
#>    27   76 0.7219315716 0.2561021829 0.7099775598 0.2366303992  0.744342647
#>    45   34 0.5791605126 0.2440093515 0.5682827467 0.2375056718  0.630105063
#>    50   63 0.8106550740 0.4585176575 0.8572706407 0.3548027258  0.845429391
#>    74   44 0.8791066229 0.7605102239 0.9204803349 0.8468039082  0.819442731
#>    74    4 0.8791066229 0.7605102239 0.8173168515 0.7410604788  0.858951635
#>    50    3 0.8106550740 0.4585176575 0.8775753305 0.3632636358  0.883937319
#>    68   50 0.8583518018 0.5423971861 0.8106550740 0.4585176575  0.889254451
#>    95   97 0.7272713816 0.1981375014 0.7278531507 0.0914161440  0.777367762
#>    57   96 0.3201244288 0.0930332358 0.1750982081 0.0982790695  0.239352882
#>    96   98 0.1750982081 0.0982790695 0.1171151995 0.1397658550  0.121856308
#>    46   42 0.0013167037 0.2621080528 0.0816309661 0.0478530301 -0.030810720
#>    99    4 0.6725360499 0.8211587877 0.8173168515 0.7410604788  0.781775569
#>    63   27 0.8572706407 0.3548027258 0.7219315716 0.2561021829  0.832125798
#>    95   34 0.7272713816 0.1981375014 0.5682827467 0.2375056718  0.629557704
#>    57   55 0.3201244288 0.0930332358 0.7257551286 0.0813661337  0.521372704
#>    49   69 0.7801595300 0.0386571041 0.8492956542 0.1326190822  0.791801114
#>    57   42 0.3201244288 0.0930332358 0.0816309661 0.0478530301  0.405310845
#>    44   70 0.9204803349 0.8468039082 0.8165272772 0.9546310835  0.813649120
#>    68   20 0.8583518018 0.5423971861 0.9595427899 0.4325248044  0.987558215
#>    60   20 0.9391770667 0.3602803897 0.9595427899 0.4325248044  0.910582668
#>    64   27 0.9422488224 0.2569385526 0.7219315716 0.2561021829  0.832143978
#>    88   65 0.9589967776 0.7437645283 0.9899584912 0.7517215705  0.967290975
#>    83   87 0.8303568039 0.6036763120 1.0000000000 0.6896468422  0.908238147
#>    88   87 0.9589967776 0.7437645283 1.0000000000 0.6896468422  0.981972387
#>    64   60 0.9422488224 0.2569385526 0.9391770667 0.3602803897  0.905751319
#>    44   30 0.9204803349 0.8468039082 0.9852659815 0.9552564829  0.967235951
#>    65   44 0.9899584912 0.7517215705 0.9204803349 0.8468039082  0.976272096
#>    87   20 1.0000000000 0.6896468422 0.9595427899 0.4325248044  2.387504991
#>     3   63 0.8775753305 0.3632636358 0.8572706407 0.3548027258  0.892281516
#>    49   67 0.7801595300 0.0386571041 0.8562244459 0.0001359739  0.841691140
#>    55   49 0.7257551286 0.0813661337 0.7801595300 0.0386571041  0.767067592
#>    56   70 0.9945216505 1.0000000000 0.8165272772 0.9546310835  0.890927530
#>    73   56 0.9782199310 0.9603509728 0.9945216505 1.0000000000  0.911835726
#>    30   73 0.9852659815 0.9552564829 0.9782199310 0.9603509728  0.955954387
#>    65   30 0.9899584912 0.7517215705 0.9852659815 0.9552564829  1.030791610
#>    87   56 1.0000000000 0.6896468422 0.9945216505 1.0000000000  1.856102546
#>    92   64 0.9932269798 0.0175548342 0.9422488224 0.2569385526  0.971614110
#>    49   57 0.7801595300 0.0386571041 0.3201244288 0.0930332358  0.461368885
#>    56   17 0.9945216505 1.0000000000 0.6264319693 0.9769434109  0.780435239
#>    56   30 0.9945216505 1.0000000000 0.9852659815 0.9552564829  1.529598833
#>    72   92 0.9311793862 0.0000000000 0.9932269798 0.0175548342  0.937644460
#>    67   72 0.8562244459 0.0001359739 0.9311793862 0.0000000000  0.893826103
#>    20   64 0.9595427899 0.4325248044 0.9422488224 0.2569385526  1.628605209
#>     1    7 0.3054776388 0.5572361150 0.4636868728 0.4278440143  0.369376830
#>     1   29 0.3054776388 0.5572361150 0.2650356960 0.4624196007  0.254680986
#>     1   36 0.3054776388 0.5572361150 0.1936345378 0.5287609463  0.251782726
#>    23   62 0.4230200637 0.7034241702 0.3634317382 0.5959901301  0.381211523
#>    53    7 0.5477162236 0.5756295718 0.4636868728 0.4278440143  0.449797151
#>     1   62 0.3054776388 0.5572361150 0.3634317382 0.5959901301  0.387934008
#>    85   15 0.5002122045 0.6716518611 0.5030238466 0.7325563598  0.469099638
#>    62   85 0.3634317382 0.5959901301 0.5002122045 0.6716518611  0.438252171
#>    43   12 0.2323059930 0.7544823097 0.3761285772 0.7353842626  0.296376211
#>    12   75 0.3761285772 0.7353842626 0.4935387162 0.8213753645  0.437390875
#>    53   90 0.5477162236 0.5756295718 0.6492330841 0.5583215095  0.598312304
#>    53   62 0.5477162236 0.5756295718 0.3634317382 0.5959901301  0.453318769
#>    86   36 0.0491712449 0.5941819538 0.1936345378 0.5287609463  0.132641050
#>    43   79 0.2323059930 0.7544823097 0.3444601990 0.6052700307  0.266834402
#>    86   66 0.0491712449 0.5941819538 0.2163473860 0.5806824935  0.132672500
#>    81   29 0.3056233310 0.2556777984 0.2650356960 0.4624196007  0.348851188
#>    71   26 0.0855045795 0.7286560312 0.2141102154 0.8078739020  0.155968412
#>    26   77 0.2141102154 0.8078739020 0.3371242858 0.7768486278  0.277459828
#>    53    2 0.5477162236 0.5756295718 0.5162246064 0.6365204073  0.612216741
#>    66   71 0.2163473860 0.5806824935 0.0855045795 0.7286560312  0.171061082
#>     2    5 0.5162246064 0.6365204073 0.5713509127 0.7741688117  0.626665694
#>    53   21 0.5477162236 0.5756295718 0.6092004780 0.3947836554  0.551120271
#>    37   16 0.1000600514 0.3537818960 0.0911648673 0.5134230284  0.121541127
#>    47   18 0.0009797839 0.6063623888 0.0222226822 0.6822887040  0.034584755
#>     5   15 0.5713509127 0.7741688117 0.5030238466 0.7325563598  0.521314508
#>     2   13 0.5162246064 0.6365204073 0.5131683391 0.7104583822  0.613165738
#>    19   39 0.6353272900 0.5302164045 0.7013501159 0.3908462874  0.657958574
#>    53   19 0.5477162236 0.5756295718 0.6353272900 0.5302164045  0.551706546
#>    37   80 0.1000600514 0.3537818960 0.2037706969 0.4524766260  0.217108159
#>    54   78 0.3842868552 0.9086257009 0.4608184374 0.8785130097  0.397386410
#>    93   24 0.5351184728 0.8710421744 0.5866144017 0.8216949868  0.540065821
#>    47   86 0.0009797839 0.6063623888 0.0491712449 0.5941819538  0.020378088
#>    84   89 0.0000000000 0.4777148316 0.0475215113 0.5659260962 -0.003516782
#>    84   47 0.0000000000 0.4777148316 0.0009797839 0.6063623888 -0.013973419
#>    54   77 0.3842868552 0.9086257009 0.3371242858 0.7768486278  0.388136203
#>    82  100 0.7402570598 0.5510603195 0.7259117501 0.5229049673  0.694885034
#>    25   71 0.0498017708 0.7732753444 0.0855045795 0.7286560312  0.037946749
#>    71   33 0.0855045795 0.7286560312 0.1059619109 0.8614958192  0.137228789
#>    59   28 0.5180089228 0.2638303876 0.4675944762 0.3370529010  0.549247320
#>    78   91 0.4608184374 0.8785130097 0.4872104806 0.9042115395  0.497237913
#>    93   75 0.5351184728 0.8710421744 0.4935387162 0.8213753645  0.496589135
#>    38   37 0.1006092894 0.2921967258 0.1000600514 0.3537818960  0.212031341
#>    32   94 0.0325758348 0.3474112355 0.0434441415 0.5009548594  0.059514913
#>    31   90 0.8086489416 0.6595662310 0.6492330841 0.5583215095  0.700273205
#>    45   59 0.5791605126 0.2440093515 0.5180089228 0.2638303876  0.571286122
#>    34   22 0.5682827467 0.2375056718 0.4336725378 0.2467560993  0.495282836
#>   100   50 0.7259117501 0.5229049673 0.8106550740 0.4585176575  0.738991520
#>    19  100 0.6353272900 0.5302164045 0.7259117501 0.5229049673  0.675566384
#>    21   27 0.6092004780 0.3947836554 0.7219315716 0.2561021829  0.645539538
#>    81   35 0.3056233310 0.2556777984 0.3226057961 0.1460559555  0.295816299
#>    51   38 0.2494884225 0.1623540158 0.1006092894 0.2921967258  0.199875107
#>    46   61 0.0013167037 0.2621080528 0.0152284580 0.3034370101  0.053915661
#>    24   40 0.5866144017 0.8216949868 0.6629957170 0.8571073676  0.629622550
#>    98   52 0.1171151995 0.1397658550 0.1002362212 0.1711840128  0.176344928
#>    52   46 0.1002362212 0.1711840128 0.0013167037 0.2621080528  0.064706004
#>    46   84 0.0013167037 0.2621080528 0.0000000000 0.4777148316 -0.246121956
#>    70   99 0.8165272772 0.9546310835 0.6725360499 0.8211587877  0.765651405
#>    82   83 0.7402570598 0.5510603195 0.8303568039 0.6036763120  0.788896344
#>   100   11 0.7259117501 0.5229049673 0.8057000284 0.4852282456  0.745076157
#>    22   57 0.4336725378 0.2467560993 0.3201244288 0.0930332358  0.453430157
#>    96   51 0.1750982081 0.0982790695 0.2494884225 0.1623540158  0.197215262
#>    14   31 0.8157441382 0.6872455277 0.8086489416 0.6595662310  0.768746259
#>     4   10 0.8173168515 0.7410604788 0.8111275515 0.7102307245  0.798942239
#>    10   14 0.8111275515 0.7102307245 0.8157441382 0.6872455277  0.776325217
#>    95   45 0.7272713816 0.1981375014 0.5791605126 0.2440093515  0.641316151
#>    17   54 0.6264319693 0.9769434109 0.3842868552 0.9086257009 -0.169980489
#>    93   17 0.5351184728 0.8710421744 0.6264319693 0.9769434109  0.603455354
#>     4   58 0.8173168515 0.7410604788 0.7769132417 0.7309053745  0.773012893
#>    34   97 0.5682827467 0.2375056718 0.7278531507 0.0914161440  0.600695759
#>    96    8 0.1750982081 0.0982790695 0.2511959396 0.1423676212  0.246425877
#>    88   74 0.9589967776 0.7437645283 0.8791066229 0.7605102239  0.910827831
#>    87   68 1.0000000000 0.6896468422 0.8583518018 0.5423971861  0.937408364
#>    82    6 0.7402570598 0.5510603195 0.8415481020 0.5709817379  0.795025006
#>    10   74 0.8111275515 0.7102307245 0.8791066229 0.7605102239  0.864609432
#>    68   11 0.8583518018 0.5423971861 0.8057000284 0.4852282456  0.810592349
#>    69   95 0.8492956542 0.1326190822 0.7272713816 0.1981375014  0.826473244
#>    63   39 0.8572706407 0.3548027258 0.7013501159 0.3908462874  0.781702180
#>    95   76 0.7272713816 0.1981375014 0.7099775598 0.2366303992  0.744342647
#>     4   44 0.8173168515 0.7410604788 0.9204803349 0.8468039082  0.819442731
#>    20    3 0.9595427899 0.4325248044 0.8775753305 0.3632636358  0.883937319
#>    20   50 0.9595427899 0.4325248044 0.8106550740 0.4585176575  0.889254451
#>    69   97 0.8492956542 0.1326190822 0.7278531507 0.0914161440  0.777367762
#>    42   96 0.0816309661 0.0478530301 0.1750982081 0.0982790695  0.239352882
#>    52   42 0.1002362212 0.1711840128 0.0816309661 0.0478530301 -0.030810720
#>    70    4 0.8165272772 0.9546310835 0.8173168515 0.7410604788  0.781775569
#>    34   55 0.5682827467 0.2375056718 0.7257551286 0.0813661337  0.521372704
#>    67   42 0.8562244459 0.0001359739 0.0816309661 0.0478530301  0.405310845
#>    88   14 0.9589967776 0.7437645283 0.8157441382 0.6872455277  0.903045498
#>    69   27 0.8492956542 0.1326190822 0.7219315716 0.2561021829  0.832143978
#>     9   65 0.9364079020 0.7640727483 0.9899584912 0.7517215705  0.967290975
#>    14   87 0.8157441382 0.6872455277 1.0000000000 0.6896468422  0.908238147
#>    44    9 0.9204803349 0.8468039082 0.9364079020 0.7640727483  0.905351155
#>    65   87 0.9899584912 0.7517215705 1.0000000000 0.6896468422  0.981972387
#>     3   60 0.8775753305 0.3632636358 0.9391770667 0.3602803897  0.905751319
#>    48   44 0.9120442995 0.9274299461 0.9204803349 0.8468039082  0.846864560
#>    48   30 0.9120442995 0.9274299461 0.9852659815 0.9552564829  0.967235951
#>    92   20 0.9932269798 0.0175548342 0.9595427899 0.4325248044  2.387504991
#>    64   63 0.9422488224 0.2569385526 0.8572706407 0.3548027258  0.892281516
#>    69   67 0.8492956542 0.1326190822 0.8562244459 0.0001359739  0.841691140
#>    97   49 0.7278531507 0.0914161440 0.7801595300 0.0386571041  0.767067592
#>    48   70 0.9120442995 0.9274299461 0.8165272772 0.9546310835  0.890927530
#>    48   56 0.9120442995 0.9274299461 0.9945216505 1.0000000000  0.911835726
#>    48   73 0.9120442995 0.9274299461 0.9782199310 0.9603509728  0.955954387
#>    65   56 0.9899584912 0.7517215705 0.9945216505 1.0000000000  1.856102546
#>    69   64 0.8492956542 0.1326190822 0.9422488224 0.2569385526  0.971614110
#>    67   57 0.8562244459 0.0001359739 0.3201244288 0.0930332358  0.461368885
#>    69   92 0.8492956542 0.1326190822 0.9932269798 0.0175548342  0.937644460
#>    69   72 0.8492956542 0.1326190822 0.9311793862 0.0000000000  0.893826103
#>    66    1 0.2163473860 0.5806824935 0.3054776388 0.5572361150  0.274846915
#>    79   23 0.3444601990 0.6052700307 0.4230200637 0.7034241702  0.365295040
#>    43   66 0.2323059930 0.7544823097 0.2163473860 0.5806824935  0.266834402
#>    81   28 0.3056233310 0.2556777984 0.4675944762 0.3370529010  0.348851188
#>     5   75 0.5713509127 0.7741688117 0.4935387162 0.8213753645  0.521314508
#>    54   41 0.3842868552 0.9086257009 0.3287631143 0.8800476613  0.388136203
#>    25   18 0.0498017708 0.7732753444 0.0222226822 0.6822887040  0.037946749
#>    78   93 0.4608184374 0.8785130097 0.5351184728 0.8710421744  0.497237913
#>    37   32 0.1000600514 0.3537818960 0.0325758348 0.3474112355  0.068950552
#>    45   21 0.5791605126 0.2440093515 0.6092004780 0.3947836554  0.571286122
#>    81   51 0.3056233310 0.2556777984 0.2494884225 0.1623540158  0.295816299
#>    24   99 0.5866144017 0.8216949868 0.6725360499 0.8211587877  0.629622550
#>    70   40 0.8165272772 0.9546310835 0.6629957170 0.8571073676  0.765651405
#>    83   31 0.8303568039 0.6036763120 0.8086489416 0.6595662310  0.765804362
#>    17   33 0.6264319693 0.9769434109 0.1059619109 0.8614958192 -0.169980489
#>    88    9 0.9589967776 0.7437645283 0.9364079020 0.7640727483  0.910827831
#>     3   64 0.8775753305 0.3632636358 0.9422488224 0.2569385526  0.905751319
#>    92   87 0.9932269798 0.0175548342 1.0000000000 0.6896468422  2.387504991
#>          my1           mx2          my2 bp1 bp2
#>   0.56992537   0.381211523   0.65637093   0   0
#>   0.62193025   0.338926803   0.56992537   0   0
#>   0.47394821   0.352604371   0.37758343   0   0
#>   0.52286926   0.228379800   0.49456210   0   0
#>   0.53425287   0.132672500   0.58635711   0   0
#>   0.66911004   0.437390875   0.77488824   0   0
#>   0.70137595   0.412586114   0.80875605   0   0
#>   0.53352346   0.387934008   0.49663829   0   0
#>   0.37758343   0.530519948   0.38524083   0   0
#>   0.62585611   0.468245666   0.70364474   0   0
#>   0.70360531   0.545992629   0.73890449   0   0
#>   0.62219655   0.554389224   0.67513012   0   0
#>   0.68588413   0.365295040   0.66911004   0   0
#>   0.77488824   0.443670495   0.77113388   0   0
#>   0.56602331   0.680521633   0.52534792   0   0
#>   0.56539781   0.438252171   0.62219655   0   0
#>   0.58628766   0.147970020   0.48387777   0   0
#>   0.66367927   0.274846915   0.62193025   0   0
#>   0.70364474   0.443670495   0.77113388   0   0
#>   0.67513012   0.469099638   0.70360531   0   0
#>   0.48387777   0.228379800   0.49456210   0   0
#>   0.37151930   0.352604371   0.37758343   0   0
#>   0.75826292   0.277459828   0.79966701   0   0
#>   0.79966701   0.298433353   0.70137595   0   0
#>   0.64254055   0.090783543   0.65508426   0   0
#>   0.57625853   0.132641050   0.58628766   0   0
#>   0.81236631   0.412586114   0.80875605   0   0
#>   0.64757691   0.626665694   0.67215313   0   0
#>   0.67247334   0.155968412   0.75826292   0   0
#>   0.67215313   0.659724647   0.68408148   0   0
#>   0.47591217   0.530519948   0.38524083   0   0
#>   0.43310369   0.060695100   0.53248581   0   0
#>   0.43504720   0.147970020   0.48387777   0   0
#>   0.53248581   0.113353797   0.57625853   0   0
#>   0.63789515   0.090783543   0.65508426   0   0
#>   0.77942562   0.443670495   0.77113388   0   0
#>   0.93886870   0.188163371   1.22144332   0   0
#>   0.82445946   0.211692327   0.93886870   0   0
#>   0.67755968   0.545992629   0.73890449   0   0
#>   0.35239791   0.228379800   0.49456210   0   0
#>   0.45561404   0.651922759   0.31435261   0   0
#>   0.47611150   0.657958574   0.45561404   0   0
#>   0.28466909   0.374476181   0.32051464   0   0
#>   0.33462321   0.251452271   0.35239791   0   0
#>   0.82960922   0.411558375   0.81236631   0   0
#>   0.82466224   0.540114747   0.81041470   0   0
#>   0.58168697   0.113353797   0.57625853   0   0
#>   0.32051464   0.366707522   0.20901451   0   0
#>   0.53651552   0.060695100   0.53248581   0   0
#>   0.54214876   0.020378088   0.58168697   0   0
#>   0.83291983   0.283712708   0.82445946   0   0
#>   0.42281728  -0.003516782   0.53651552   0   0
#>   0.81041470   0.629327702   0.78176310   0   0
#>   0.55644544   0.680521633   0.52534792   0   0
#>   0.72719565   0.090783543   0.65508426   0   0
#>   0.78868561   0.211692327   0.93886870   0   0
#>   0.79158303  -7.947963872   5.92654670   0   1
#>   0.33930503   0.530519948   0.38524083   0   0
#>   0.86751210   0.552952419   0.94798304   0   0
#>   0.86105978   0.411558375   0.81236631   0   0
#>   0.92963802   0.446287858   1.15215762   0   0
#>   0.32398546   0.217108159   0.33462321   0   0
#>   0.32270942   0.059694493   0.31130441   0   0
#>   1.22144332  -0.169980489   3.33645755   0   0
#>   0.42266086   0.086661030   0.43310369   0   0
#>   0.65408304   0.659724647   0.68408148   0   0
#>   0.27264762   0.495282836   0.15926131   0   0
#>   0.68408148   0.664391992   0.70625796   0   0
#>   0.32395793   0.549247320   0.33930503   0   0
#>   0.15926131   0.469893456   0.28466909   0   0
#>   0.45215880   0.781702180   0.38317118   0   0
#>   0.46395526   0.738991520   0.45215880   0   0
#>   0.30916383   0.647287942   0.28852173   0   0
#>   0.20901451   0.453430157   0.11336419   0   0
#>   0.19803214   0.286321205   0.15543465   0   0
#>   0.25574143   0.212031341   0.32398546   0   0
#>   0.26740860   0.059694493   0.31130441   0   0
#>   0.88906382   0.603455354   0.90443681   0   0
#>   0.82901028   0.765651405   0.86511075   0   0
#>   0.31130441  -0.114257609   0.37992672   0   0
#>   0.37992672   0.057305003   0.42281728   0   0
#>   0.08616016   0.521372704   0.03271712   0   0
#>   0.94798304   0.446287858   1.15215762   0   0
#>   0.19182930   0.178688897   0.23144908   0   0
#>   0.23180048   0.053915661   0.26740860   0   0
#>   0.36840436  -0.114257609   0.37992672   0   0
#>   0.23144908   0.064706004   0.23180048   0   0
#>   0.70625796   0.629327702   0.78176310   0   0
#>   0.57122180   0.929090158   0.61920991   0   0
#>   0.51716055   0.795025006   0.54006044   0   0
#>   0.46016710   0.865930912   0.48258670   0   0
#>   0.11336419   0.288094514   0.12110160   0   0
#>   0.12110160   0.286321205   0.15543465   0   0
#>   0.16716613   0.176344928   0.19182930   0   0
#>   0.65072495   0.694885034   0.55644544   0   0
#>   0.14782197   0.286321205   0.15543465   0   0
#>   0.11750421   0.121856308   0.08512948   0   0
#>   0.31435261   0.769961532   0.33238241   0   0
#>   0.68454374   0.700273205   0.65408304   0   0
#>   0.61076456   0.702405926   0.65072495   0   0
#>   0.65566488   0.908238147   0.66035658   0   0
#>   0.72871317   0.776325217   0.69128444   0   0
#>   0.69128444   0.768746259   0.68454374   0   0
#>   0.18265141   0.647287942   0.28852173   0   0
#>   0.90443681   0.718394945   0.93950658   0   0
#>   0.83187697   0.664391992   0.70625796   0   0
#>   0.11271730   0.767067592   0.07798581   0   0
#>   0.06288346   0.288094514   0.12110160   0   0
#>   0.71290305   0.905351155   0.80099243   0   0
#>   0.60810272   0.810592349   0.53355272   0   0
#>   0.54006044   0.810592349   0.53355272   0   0
#>   0.93950658   0.780435239   1.46807397   0   0
#>   0.70901636   0.903045498   0.67577519   0   0
#>   0.61920991   0.937408364   0.60810272   0   0
#>   0.23650445   0.744342647   0.22893843   0   0
#>   0.22893843   0.647287942   0.28852173   0   0
#>   0.14645308   0.554654935   0.27264762   0   0
#>   0.41181392   0.781702180   0.38317118   0   0
#>   0.84218142   0.905351155   0.80099243   0   0
#>   0.71666584   0.819442731   0.84218142   0   0
#>   0.43886748   0.845429391   0.41181392   0   0
#>   0.46932415   0.865930912   0.48258670   0   0
#>   0.14504833   0.629557704   0.14424257   0   0
#>  -0.13265643   0.246425877   0.06288346   0   0
#>   0.08512948   0.180553451   0.16716613   0   0
#>   0.12788442  -4.123114907  -1.40613046   0   1
#>   0.84771584   0.773012893   0.83187697   0   0
#>   0.24714222   0.769961532   0.33238241   0   0
#>   0.14424257   0.630105063   0.14645308   0   0
#>   0.03271712   0.513379833  -0.24517139   0   0
#>   0.10250713   0.841691140   0.06579863   0   0
#>  -1.00870134   0.239352882  -0.13265643   0   0
#>   0.84783368   0.846864560   0.87985573   0   0
#>   0.55986060   0.889254451   0.46932415   0   0
#>   0.40733392   1.230262395   0.31721610   0   0
#>   0.24235341   0.832125798   0.24714222   0   0
#>   0.77570712   0.981972387   0.71858015   0   0
#>   0.66035658   0.929090158   0.61920991   0   0
#>   0.71858015   0.908207723   0.66269105   0   0
#>   0.30757026   1.230262395   0.31721610   0   0
#>   0.89245038   1.030791610   0.85448453   0   0
#>   0.81464627   1.030791610   0.85448453   0   0
#>   0.33958410   0.987558215   0.55986060   0   0
#>   0.29937710   0.845429391   0.41181392   0   0
#>   0.06579863   0.461368885  -0.68519648   0   0
#>   0.07798581   0.513379833  -0.24517139   0   0
#>   1.03458321   0.780435239   1.46807397   0   0
#>   1.01082062   0.995272206   0.97651566   0   0
#>   0.92213626   0.995272206   0.97651566   0   0
#>   0.85448453   1.529598833   0.86598456   0   0
#>   0.85998368   5.077234600   0.91684306   0   1
#>   0.13807215   1.628605209   0.27798224   0   0
#>  -0.68519648   0.513379833  -0.24517139   0   0
#>   1.46807397   0.610928442   4.17418556   0   1
#>   0.86598456   0.995272206   0.97651566   0   0
#>   0.09558027   5.159509067 -14.82660788   0   1
#>   0.06852526   0.869623965 -13.27277960   0   1
#>   0.27798224   1.230262395   0.31721610   0   0
#>   0.47394821   0.387934008   0.49663829   0   0
#>   0.52286926   0.369376830   0.47394821   0   0
#>   0.53425287   0.254680986   0.52286926   0   0
#>   0.65637093   0.436227842   0.62585611   0   0
#>   0.53352346   0.551120271   0.47591217   0   0
#>   0.49663829   0.338926803   0.56992537   0   0
#>   0.70360531   0.468245666   0.70364474   0   0
#>   0.62219655   0.436227842   0.62585611   0   0
#>   0.68588413   0.298433353   0.70137595   0   0
#>   0.77488824   0.412586114   0.80875605   0   0
#>   0.56602331   0.612216741   0.64757691   0   0
#>   0.56539781   0.449797151   0.53352346   0   0
#>   0.58628766   0.132672500   0.58635711   0   0
#>   0.66367927   0.296376211   0.68588413   0   0
#>   0.58635711   0.137209309   0.64254055   0   0
#>   0.37151930   0.251452271   0.35239791   0   0
#>   0.75826292   0.137228789   0.78868561   0   0
#>   0.79966701   0.283712708   0.82445946   0   0
#>   0.64757691   0.453318769   0.56539781   0   0
#>   0.67247334   0.137209309   0.64254055   0   0
#>   0.67215313   0.613165738   0.67755968   0   0
#>   0.47591217   0.551706546   0.47611150   0   0
#>   0.43504720   0.086661030   0.43310369   0   0
#>   0.63789515 -12.179943592   4.05531351   0   1
#>   0.77942562   0.545992629   0.73890449   0   0
#>   0.67755968   0.554389224   0.67513012   0   0
#>   0.45561404   0.675566384   0.46395526   0   0
#>   0.47611150   0.598312304   0.56602331   0   0
#>   0.33462321   0.121541127   0.43504720   0   0
#>   0.82960922   0.436744496   0.92963802   0   0
#>   0.82466224   0.601780156   0.88906382   0   0
#>   0.58168697   0.034584755   0.63789515   0   0
#>   0.53651552  -0.013973419   0.54214876   0   0
#>   0.54214876  -7.786697759   0.60134608   0   1
#>   0.83291983   0.397386410   0.82960922   0   0
#>   0.55644544   0.771988965   0.51716055   0   0
#>   0.72719565   0.118414387   0.79158303   0   0
#>   0.78868561   0.118414387   0.79158303   0   0
#>   0.33930503   0.469893456   0.28466909   0   0
#>   0.86751210   0.436744496   0.92963802   0   0
#>   0.86105978   0.540065821   0.82466224   0   0
#>   0.32398546   0.068950552   0.32270942   0   0
#>   0.42266086   0.057305003   0.42281728   0   0
#>   0.65408304   0.702405926   0.65072495   0   0
#>   0.32395793   0.554654935   0.27264762   0   0
#>   0.15926131   0.490259318   0.08616016   0   0
#>   0.45215880   0.745076157   0.46016710   0   0
#>   0.46395526   0.680521633   0.52534792   0   0
#>   0.30916383   0.651922759   0.31435261   0   0
#>   0.19803214   0.366707522   0.20901451   0   0
#>   0.25574143   0.178688897   0.23144908   0   0
#>   0.26740860  -0.246121956   0.36840436   0   0
#>   0.82901028   0.601780156   0.88906382   0   0
#>   0.19182930   0.037997847   0.11750421   0   0
#>   0.23180048  -0.030810720   0.12788442   0   0
#>   0.36840436  -4.884022083   0.34008085   0   1
#>   0.86511075   0.781775569   0.84771584   0   0
#>   0.57122180   0.765804362   0.61076456   0   0
#>   0.46016710   0.771988965   0.51716055   0   0
#>   0.11336419   0.490259318   0.08616016   0   0
#>   0.14782197   0.180553451   0.16716613   0   0
#>   0.68454374   0.881406504   0.65566488   0   0
#>   0.72871317   0.858951635   0.71666584   0   0
#>   0.69128444   0.864609432   0.70901636   0   0
#>   0.18265141   0.630105063   0.14645308   0   0
#>   3.33645755   0.446287858   1.15215762   0   0
#>   0.90443681   0.552952419   0.94798304   0   0
#>   0.83187697   0.798942239   0.72871317   0   0
#>   0.11271730   0.629557704   0.14424257   0   0
#>   0.06288346   0.197215262   0.14782197   0   0
#>   0.71290305   0.903045498   0.67577519   0   0
#>   0.60810272   0.987558215   0.55986060   0   0
#>   0.54006044   0.788896344   0.57122180   0   0
#>   0.70901636   0.858951635   0.71666584   0   0
#>   0.53355272   0.865930912   0.48258670   0   0
#>   0.23650445   0.777367762   0.14504833   0   0
#>   0.38317118   0.769961532   0.33238241   0   0
#>   0.22893843   0.641316151   0.18265141   0   0
#>   0.84218142   0.813649120   0.84783368   0   0
#>   0.43886748   0.910582668   0.40733392   0   0
#>   0.46932415   0.883937319   0.43886748   0   0
#>   0.14504833   0.791801114   0.10250713   0   0
#>  -0.13265643   0.121856308   0.08512948   0   0
#>   0.12788442   0.037997847   0.11750421   0   0
#>   0.84771584   0.813649120   0.84783368   0   0
#>   0.03271712   0.600695759   0.11271730   0   0
#>  -1.00870134   0.326082427  -2.29482042   0   1
#>   0.67577519   0.908207723   0.66269105   0   0
#>   0.24235341   0.826473244   0.23650445   0   0
#>   0.77570712   0.976272096   0.81464627   0   0
#>   0.66035658   0.908207723   0.66269105   0   0
#>   0.80099243   0.976272096   0.81464627   0   0
#>   0.71858015   1.856102546   0.85998368   0   0
#>   0.30757026   0.910582668   0.40733392   0   0
#>   0.87985573   0.967235951   0.89245038   0   0
#>   0.89245038   0.955954387   0.92213626   0   0
#>   0.33958410   1.628605209   0.27798224   0   0
#>   0.29937710   0.832125798   0.24714222   0   0
#>   0.06579863   0.893826103   0.06852526   0   0
#>   0.07798581   0.791801114   0.10250713   0   0
#>   1.03458321   0.846864560   0.87985573   0   0
#>   1.01082062   0.890927530   1.03458321   0   0
#>   0.92213626   0.911835726   1.01082062   0   0
#>   0.85998368   1.529598833   0.86598456   0   0
#>   0.13807215   0.832143978   0.24235341   0   0
#>  -0.68519648   0.405310845  -1.00870134   0   0
#>   0.09558027   0.971614110   0.13807215   0   0
#>   0.06852526   0.937644460   0.09558027   0   0
#>   0.62193025   0.251782726   0.53425287   0   0
#>   0.66911004   0.381211523   0.65637093   0   0
#>   0.66367927   0.171061082   0.67247334   0   0
#>   0.37151930   0.374476181   0.32051464   0   0
#>   0.77942562   0.540114747   0.81041470   0   0
#>   0.83291983   0.188163371   1.22144332   0   0
#>   0.72719565 -10.027865879   3.77825827   0   1
#>   0.86751210   0.496589135   0.86105978   0   0
#>   0.32270942   0.059514913   0.42266086   0   0
#>   0.32395793   0.645539538   0.30916383   0   0
#>   0.19803214   0.199875107   0.25574143   0   0
#>   0.82901028   0.629327702   0.78176310   0   0
#>   0.86511075   0.718394945   0.93950658   0   0
#>   0.61076456   0.881406504   0.65566488   0   0
#>   3.33645755  -0.576175400   5.16769791   0   1
#>   0.71290305   0.967290975   0.77570712   0   0
#>   0.30757026   0.892281516   0.29937710   0   0
#>   0.33958410   3.875245422   0.32459137   0   1
#> 
#> $x
#>                [,1]         [,2]
#>   [1,] 0.3054776388 0.5572361150
#>   [2,] 0.5162246064 0.6365204073
#>   [3,] 0.8775753305 0.3632636358
#>   [4,] 0.8173168515 0.7410604788
#>   [5,] 0.5713509127 0.7741688117
#>   [6,] 0.8415481020 0.5709817379
#>   [7,] 0.4636868728 0.4278440143
#>   [8,] 0.2511959396 0.1423676212
#>   [9,] 0.9364079020 0.7640727483
#>  [10,] 0.8111275515 0.7102307245
#>  [11,] 0.8057000284 0.4852282456
#>  [12,] 0.3761285772 0.7353842626
#>  [13,] 0.5131683391 0.7104583822
#>  [14,] 0.8157441382 0.6872455277
#>  [15,] 0.5030238466 0.7325563598
#>  [16,] 0.0911648673 0.5134230284
#>  [17,] 0.6264319693 0.9769434109
#>  [18,] 0.0222226822 0.6822887040
#>  [19,] 0.6353272900 0.5302164045
#>  [20,] 0.9595427899 0.4325248044
#>  [21,] 0.6092004780 0.3947836554
#>  [22,] 0.4336725378 0.2467560993
#>  [23,] 0.4230200637 0.7034241702
#>  [24,] 0.5866144017 0.8216949868
#>  [25,] 0.0498017708 0.7732753444
#>  [26,] 0.2141102154 0.8078739020
#>  [27,] 0.7219315716 0.2561021829
#>  [28,] 0.4675944762 0.3370529010
#>  [29,] 0.2650356960 0.4624196007
#>  [30,] 0.9852659815 0.9552564829
#>  [31,] 0.8086489416 0.6595662310
#>  [32,] 0.0325758348 0.3474112355
#>  [33,] 0.1059619109 0.8614958192
#>  [34,] 0.5682827467 0.2375056718
#>  [35,] 0.3226057961 0.1460559555
#>  [36,] 0.1936345378 0.5287609463
#>  [37,] 0.1000600514 0.3537818960
#>  [38,] 0.1006092894 0.2921967258
#>  [39,] 0.7013501159 0.3908462874
#>  [40,] 0.6629957170 0.8571073676
#>  [41,] 0.3287631143 0.8800476613
#>  [42,] 0.0816309661 0.0478530301
#>  [43,] 0.2323059930 0.7544823097
#>  [44,] 0.9204803349 0.8468039082
#>  [45,] 0.5791605126 0.2440093515
#>  [46,] 0.0013167037 0.2621080528
#>  [47,] 0.0009797839 0.6063623888
#>  [48,] 0.9120442995 0.9274299461
#>  [49,] 0.7801595300 0.0386571041
#>  [50,] 0.8106550740 0.4585176575
#>  [51,] 0.2494884225 0.1623540158
#>  [52,] 0.1002362212 0.1711840128
#>  [53,] 0.5477162236 0.5756295718
#>  [54,] 0.3842868552 0.9086257009
#>  [55,] 0.7257551286 0.0813661337
#>  [56,] 0.9945216505 1.0000000000
#>  [57,] 0.3201244288 0.0930332358
#>  [58,] 0.7769132417 0.7309053745
#>  [59,] 0.5180089228 0.2638303876
#>  [60,] 0.9391770667 0.3602803897
#>  [61,] 0.0152284580 0.3034370101
#>  [62,] 0.3634317382 0.5959901301
#>  [63,] 0.8572706407 0.3548027258
#>  [64,] 0.9422488224 0.2569385526
#>  [65,] 0.9899584912 0.7517215705
#>  [66,] 0.2163473860 0.5806824935
#>  [67,] 0.8562244459 0.0001359739
#>  [68,] 0.8583518018 0.5423971861
#>  [69,] 0.8492956542 0.1326190822
#>  [70,] 0.8165272772 0.9546310835
#>  [71,] 0.0855045795 0.7286560312
#>  [72,] 0.9311793862 0.0000000000
#>  [73,] 0.9782199310 0.9603509728
#>  [74,] 0.8791066229 0.7605102239
#>  [75,] 0.4935387162 0.8213753645
#>  [76,] 0.7099775598 0.2366303992
#>  [77,] 0.3371242858 0.7768486278
#>  [78,] 0.4608184374 0.8785130097
#>  [79,] 0.3444601990 0.6052700307
#>  [80,] 0.2037706969 0.4524766260
#>  [81,] 0.3056233310 0.2556777984
#>  [82,] 0.7402570598 0.5510603195
#>  [83,] 0.8303568039 0.6036763120
#>  [84,] 0.0000000000 0.4777148316
#>  [85,] 0.5002122045 0.6716518611
#>  [86,] 0.0491712449 0.5941819538
#>  [87,] 1.0000000000 0.6896468422
#>  [88,] 0.9589967776 0.7437645283
#>  [89,] 0.0475215113 0.5659260962
#>  [90,] 0.6492330841 0.5583215095
#>  [91,] 0.4872104806 0.9042115395
#>  [92,] 0.9932269798 0.0175548342
#>  [93,] 0.5351184728 0.8710421744
#>  [94,] 0.0434441415 0.5009548594
#>  [95,] 0.7272713816 0.1981375014
#>  [96,] 0.1750982081 0.0982790695
#>  [97,] 0.7278531507 0.0914161440
#>  [98,] 0.1171151995 0.1397658550
#>  [99,] 0.6725360499 0.8211587877
#> [100,] 0.7259117501 0.5229049673
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,62,79), [7,2,11]
#> 2: (66,1,79), [1,21,5]
#> 3: (1,29,7), [12,11,4]
#> 4: (1,36,29), [9,3,5]
#> 5: (1,66,36), [26,4,2]
#> 6: (79,23,12), [17,16,7]
#> 7: (23,79,62), [1,13,6]
#> 8: (43,12,77), [34,30,16]
#> 9: (80,29,36), [4,25,48]
#> 10: (53,62,7), [11,39,19]
#> 11: (1,7,62), [10,1,3]
#> 12: (29,28,7), [27,3,28]
#> 13: (62,85,23), [22,7,15]
#> 14: (85,13,15), [35,22,23]
#> 15: (62,2,85), [23,13,19]
#> 16: (43,79,12), [6,8,21]
#> 17: (12,23,75), [24,34,6]
#> 18: (53,19,90), [76,36,50]
#> 19: (53,2,62), [15,10,36]
#> 20: (86,16,36), [25,26,32]
#> 21: (43,66,79), [2,16,37]
#> 22: (85,15,23), [24,13,14]
#> 23: (2,13,85), [14,15,47]
#> 24: (75,23,15), [22,44,17]
#> 25: (16,80,36), [9,20,41]
#> 26: (86,36,66), [5,31,20]
#> 27: (21,7,28), [12,67,39]
#> 28: (81,28,29), [12,48,57]
#> 29: (71,43,26), [30,65,37]
#> 30: (26,43,77), [8,46,29]
#> 31: (66,71,86), [56,26,37]
#> 32: (89,16,86), [20,55,42]
#> 33: (78,77,75), [34,69,53]
#> 34: (12,75,77), [33,8,17]
#> 35: (5,15,13), [14,47,44]
#> 36: (53,90,2), [38,19,18]
#> 37: (66,43,71), [29,31,21]
#> 38: (2,90,5), [78,47,36]
#> 39: (53,7,21), [27,50,10]
#> 40: (37,16,94), [42,74,41]
#> 41: (37,80,16), [25,40,52]
#> 42: (94,16,89), [32,58,40]
#> 43: (47,86,18), [56,0,55]
#> 44: (5,75,15), [24,35,62]
#> 45: (26,41,33), [73,65,46]
#> 46: (77,41,26), [45,30,60]
#> 47: (2,5,13), [35,23,38]
#> 48: (81,29,80), [9,52,28]
#> 49: (19,21,39), [112,82,50]
#> 50: (53,21,19), [49,18,39]
#> 51: (59,28,22), [57,80,67]
#> 52: (37,81,80), [48,41,71]
#> 53: (54,77,78), [33,70,60]
#> 54: (93,75,24), [62,88,69]
#> 55: (47,89,86), [32,43,59]
#> 56: (18,86,71), [31,64,43]
#> 57: (28,81,22), [84,51,28]
#> 58: (84,94,89), [42,59,61]
#> 59: (84,89,47), [55,0,58]
#> 60: (54,41,77), [46,53,73]
#> 61: (32,94,84), [58,92,74]
#> 62: (75,5,24), [90,54,44]
#> 63: (82,90,100), [76,103,108]
#> 64: (25,18,71), [56,66,0]
#> 65: (71,26,33), [45,66,29]
#> 66: (71,33,25), [0,64,65]
#> 67: (59,21,28), [27,51,79]
#> 68: (78,93,91), [94,70,69]
#> 69: (93,78,75), [33,54,68]
#> 70: (78,91,54), [122,53,68]
#> 71: (38,81,37), [52,72,86]
#> 72: (37,32,38), [91,71,74]
#> 73: (41,54,33), [121,45,60]
#> 74: (32,37,94), [40,61,72]
#> 75: (31,58,90), [78,108,113]
#> 76: (19,100,90), [63,18,82]
#> 77: (45,59,34), [80,137,79]
#> 78: (90,58,5), [101,38,75]
#> 79: (45,21,59), [67,77,83]
#> 80: (34,59,22), [51,93,77]
#> 81: (100,39,50), [135,104,82]
#> 82: (19,39,100), [81,76,49]
#> 83: (21,45,27), [119,112,79]
#> 84: (81,35,22), [105,57,85]
#> 85: (81,51,35), [95,84,86]
#> 86: (51,81,38), [71,99,85]
#> 87: (46,38,61), [91,98,97]
#> 88: (24,40,93), [123,54,89]
#> 89: (24,99,40), [100,88,90]
#> 90: (99,24,5), [62,101,89]
#> 91: (38,32,61), [92,87,72]
#> 92: (61,32,84), [61,98,91]
#> 93: (22,57,34), [150,80,105]
#> 94: (93,17,91), [122,68,123]
#> 95: (8,35,51), [85,109,106]
#> 96: (98,51,52), [99,110,107]
#> 97: (52,38,46), [87,146,99]
#> 98: (46,61,84), [92,0,87]
#> 99: (51,38,52), [97,96,86]
#> 100: (70,40,99), [89,147,130]
#> 101: (58,99,5), [90,78,124]
#> 102: (82,6,83), [132,114,129]
#> 103: (100,11,82), [129,63,104]
#> 104: (100,50,11), [111,103,81]
#> 105: (22,35,57), [106,93,84]
#> 106: (57,35,8), [95,126,105]
#> 107: (96,51,98), [96,145,109]
#> 108: (31,90,82), [63,114,75]
#> 109: (96,8,51), [95,107,126]
#> 110: (52,42,98), [145,96,146]
#> 111: (68,11,50), [104,142,133]
#> 112: (21,27,39), [118,49,83]
#> 113: (14,58,31), [75,115,117]
#> 114: (83,31,82), [108,102,115]
#> 115: (31,83,14), [159,113,114]
#> 116: (4,58,10), [117,140,124]
#> 117: (10,58,14), [113,131,116]
#> 118: (63,39,27), [112,148,135]
#> 119: (76,27,45), [83,120,136]
#> 120: (95,76,45), [119,137,136]
#> 121: (17,33,54), [73,122,0]
#> 122: (17,54,91), [70,94,121]
#> 123: (93,40,17), [130,94,88]
#> 124: (4,99,58), [101,116,147]
#> 125: (34,55,97), [171,149,150]
#> 126: (96,57,8), [106,109,144]
#> 127: (88,9,74), [160,154,158]
#> 128: (87,6,68), [133,155,132]
#> 129: (82,11,6), [133,102,103]
#> 130: (40,70,17), [180,123,100]
#> 131: (10,14,74), [154,140,117]
#> 132: (83,6,87), [128,159,102]
#> 133: (68,6,11), [129,111,128]
#> 134: (69,27,95), [136,143,157]
#> 135: (63,50,39), [81,118,138]
#> 136: (95,27,76), [119,120,134]
#> 137: (95,45,34), [77,149,120]
#> 138: (3,50,63), [135,169,141]
#> 139: (4,74,44), [160,153,140]
#> 140: (10,74,4), [139,116,131]
#> 141: (20,50,3), [138,156,142]
#> 142: (20,68,50), [111,141,155]
#> 143: (69,95,97), [149,151,134]
#> 144: (42,57,96), [126,145,152]
#> 145: (42,96,98), [107,110,144]
#> 146: (52,46,42), [0,110,97]
#> 147: (70,99,4), [124,153,100]
#> 148: (64,63,27), [118,157,169]
#> 149: (97,95,34), [137,125,143]
#> 150: (34,57,55), [172,125,93]
#> 151: (97,49,69), [170,143,171]
#> 152: (67,57,42), [144,0,179]
#> 153: (4,44,70), [164,147,139]
#> 154: (88,74,14), [131,162,127]
#> 155: (87,68,20), [142,167,128]
#> 156: (3,60,20), [168,141,163]
#> 157: (69,64,27), [148,134,178]
#> 158: (9,88,65), [161,166,127]
#> 159: (14,83,87), [132,162,115]
#> 160: (44,74,9), [127,166,139]
#> 161: (65,88,87), [162,177,158]
#> 162: (14,87,88), [161,154,159]
#> 163: (3,64,60), [168,156,169]
#> 164: (48,70,44), [153,165,173]
#> 165: (48,44,30), [176,175,164]
#> 166: (9,65,44), [176,160,158]
#> 167: (92,87,20), [155,185,0]
#> 168: (64,20,60), [156,163,185]
#> 169: (64,3,63), [138,148,163]
#> 170: (69,49,67), [179,184,151]
#> 171: (97,55,49), [172,151,125]
#> 172: (57,49,55), [171,150,179]
#> 173: (48,56,70), [180,164,174]
#> 174: (48,73,56), [181,173,175]
#> 175: (48,30,73), [181,174,165]
#> 176: (44,65,30), [182,165,166]
#> 177: (65,87,56), [0,182,161]
#> 178: (69,92,64), [185,157,183]
#> 179: (67,49,57), [172,152,170]
#> 180: (70,56,17), [0,130,173]
#> 181: (56,73,30), [175,182,174]
#> 182: (65,56,30), [181,176,177]
#> 183: (69,72,92), [0,178,184]
#> 184: (69,67,72), [0,183,170]
#> 185: (92,20,64), [168,178,167]
#> boundary nodes:  18 47 84 46 42 67 72 92 87 56 17 33 25 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.021119561 0.061861919 0.201637698 0.097464350 0.056672044 0.056747326 
#>                                                                         
#> 0.056926517 0.195765697 0.090875165 0.083475183 0.024315249 0.038608497 
#>                                                                         
#> 0.133912656 0.137424032 0.031357105 0.158077082 0.103611221 0.130450910 
#>                                                                         
#> 0.085142761 0.040912193 0.076954787 0.238216013 0.056406989 0.107178014 
#>                                                                         
#> 0.139296047 0.091024548 0.162628731 0.154292977 0.149055868 0.229468278 
#>                                                                         
#> 0.149221971 0.049322641 0.128041153 0.068273821 0.092135902 0.089324035 
#>                                                                         
#> 0.223572241 0.135478186 0.086279955 0.062066600 0.092233717 0.137929837 
#>                                                                         
#> 0.096458335 0.221593632 0.160112116 0.093076234 0.028303977 0.128359636 
#>                                                                         
#> 0.065099052 0.061654085 0.103537190 0.049269588 0.049917045 0.084462709 
#>                                                                         
#> 0.078450797 0.120711912 0.104579221 0.152921875 0.058269911 0.065843201 
#>                                                                         
#> 0.103018239 0.227773299 0.087619593 0.282287089 0.157687166 0.214679792 
#>                                                                         
#> 0.056748991 0.210065695 0.159576496 0.086047412 0.128557334 0.134323374 
#>                                                                         
#> 0.143282280 0.149921113 0.074911795 0.208241181 0.086117538 0.128634243 
#>                                                                         
#> 0.037192988 0.047272233 0.134313857 0.287149500 0.157074976 0.149513171 
#>                                                                         
#> 0.103751365 0.174941890 0.121013288 0.111563838 0.034556914 0.092825881 
#>                                                                         
#> 0.027166303 0.053080750 0.071505045 0.134286616 0.091313137 0.020059202 
#>                                                                         
#> 0.098524607 0.136306896 0.078079626 0.128261383 0.084837161 0.039975744 
#>                                                                         
#> 0.058429622 0.131024993 0.125290003 0.137986509 0.010266665 0.084764487 
#>                                                                         
#> 0.057411917 0.033157819 0.092944854 0.191400277 0.096863410 0.197960630 
#>                                                                         
#> 0.058210118 0.022848386 0.012673738 0.113709270 0.095699446 0.064778611 
#>                                                                         
#> 0.116411552 0.096492244 0.106722943 0.145121065 0.071296442 0.228813451 
#>                                                                         
#> 0.165460629 0.167506599 0.163790228 0.405798455 0.116656149 0.242735211 
#>                                                                         
#> 0.149776293 0.149370534 0.075060097 0.220318838 0.031967831 0.190183454 
#>                                                                         
#> 0.067896894 0.103387480 0.126329494 0.117761900 0.260285474 0.021996987 
#>                                                                         
#> 0.085262823 0.069165744 0.183685425 0.042869470 0.008694864 0.203588998 
#>                                                                         
#> 0.310401506 0.244751583 0.463237583 0.368811089 0.045690806 0.064483146 
#>                                                                         
#> 0.074955064 0.176435861 0.204383163 0.103081143 0.115411067 0.122852926 
#>                                                                         
#> 0.170004420 0.069717654 0.060969364 0.156312487 0.145085048 0.145532163 
#>                                                                         
#> 0.102981756 0.185405836 0.158586100 0.186662450 0.167720296 0.210688227 
#>                                                                         
#> 0.151045955 0.126866186 0.068552285 0.197524702 0.148276744 0.191011934 
#>                                                                         
#> 0.159888760 0.078842032 0.080001151 0.074001114 0.154217519 0.098681582 
#>                                                                         
#> 0.143166154 0.082242673 0.071323037 0.049706940 0.100197411 0.128651288 
#>                                                                         
#> 0.139962513 0.031599237 0.057145198 0.134405773 0.088899679 0.036836862 
#>                                                                         
#> 0.064773978 0.061587619 0.153927790 0.188848906 0.064283671 0.134927680 
#>                                                                         
#> 0.106429115 0.090879049 0.178720033 0.110929494 0.197545250 0.043607564 
#>                                                                         
#> 0.084191104 0.035665117 0.134358669 0.215610799 0.196337279 0.104337944 
#>                                                                         
#> 0.088236640 0.191112249 0.098180969 0.028574207 0.031444891 0.023444235 
#>                                                                         
#> 0.155051785 0.251598024 0.139832852 0.041660267 0.216344318 0.087946945 
#>                                                                         
#> 0.081626314 0.204320027 0.103231478 0.084552837 0.077720634 0.138501214 
#>                                                                         
#> 0.160032335 0.042199283 0.147730759 0.107311574 0.151139606 0.128241817 
#>                                                                         
#> 0.106202216 0.124726448 0.213572064 0.221759118 0.776061838 0.153999078 
#>                                                                         
#> 0.177396972 0.054956503 0.184271509 0.084250414 0.062881665 0.061673930 
#>                                                                         
#> 0.081066175 0.078330906 0.416334842 0.129609752 0.132664170 0.074293160 
#>                                                                         
#> 0.099314669 0.109858664 0.073912166 0.248320360 0.155227646 0.544089266 
#>                                                                         
#> 0.184271560 0.155861370 0.092162545 0.125721468 0.174530952 0.181263783 
#>                                                                         
#> 0.091012068 0.062446698 0.095074575 0.074674685 0.067784252 0.153737732 
#>                                                                         
#> 0.108905722 0.085923321 0.181886820 0.059957604 0.533120275 0.030375666 
#>                         
#> 0.124449523 0.672126135 
#> 
#> $alpha
#> [1] 0.2812922
#> 
#> attr(,"class")
#> [1] "scree"