Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2         mx1
#>    55   12 0.437077051 0.741109322 0.437636470 0.922814087  0.37279121
#>    12   10 0.437636470 0.922814087 0.289537368 0.922974834  0.36349036
#>    10   89 0.289537368 0.922974834 0.340470391 0.701481744  0.33091589
#>    89   66 0.340470391 0.701481744 0.365108852 0.698697555  0.35786264
#>    66   80 0.365108852 0.698697555 0.369626749 0.575333603  0.43494127
#>    80   55 0.369626749 0.575333603 0.437077051 0.741109322  0.46372399
#>    71   49 0.727295976 0.461330952 0.763411248 0.659759006  0.72537073
#>    49   63 0.763411248 0.659759006 0.700295871 0.740771609  0.73934968
#>    63   31 0.700295871 0.740771609 0.647576029 0.638622528  0.65589100
#>    31   71 0.647576029 0.638622528 0.727295976 0.461330952  0.71719200
#>    81   32 0.863912142 0.107388352 0.866843740 0.088676758  0.89607714
#>    32   90 0.866843740 0.088676758 0.974688821 0.049237969  0.93211633
#>    90   21 0.974688821 0.049237969 0.916110734 0.243279064  0.96866078
#>    21   94 0.916110734 0.243279064 0.750937479 0.303513194  0.81165429
#>    94   35 0.750937479 0.303513194 0.680527320 0.216714241  0.79036619
#>    35   18 0.680527320 0.216714241 0.769697119 0.073718991  0.78031068
#>    18   81 0.769697119 0.073718991 0.863912142 0.107388352  0.79873234
#>    93   72 0.251334318 0.421545532 0.089909767 0.352175217  0.18390234
#>    72   59 0.089909767 0.352175217 0.090963338 0.171797280  0.12660684
#>    59   87 0.090963338 0.171797280 0.219389449 0.196444365  0.14489883
#>    87   24 0.219389449 0.196444365 0.299981229 0.180103353  0.26834205
#>    24   17 0.299981229 0.180103353 0.329690249 0.393045069  0.33133922
#>    17   50 0.329690249 0.393045069 0.308579621 0.409562934  0.26953314
#>    50   93 0.308579621 0.409562934 0.251334318 0.421545532  0.26436721
#>    24   56 0.299981229 0.180103353 0.507010626 0.359129269  0.43218770
#>    56   34 0.507010626 0.359129269 0.520841007 0.384261149  0.48641950
#>    34   11 0.520841007 0.384261149 0.453187123 0.424849812  0.48520426
#>    11   17 0.453187123 0.424849812 0.329690249 0.393045069  0.39319268
#>    57   30 0.008205552 0.453984409 0.020345608 0.424632565  0.03906091
#>    30   69 0.020345608 0.424632565 0.068199061 0.427607207  0.04329629
#>    69   85 0.068199061 0.427607207 0.136806574 0.475124385  0.11298012
#>    85   46 0.136806574 0.475124385 0.107611584 0.602059228  0.08063757
#>    46   57 0.107611584 0.602059228 0.008205552 0.453984409  0.04207416
#>    48   47 0.854226231 0.994574692 0.392566529 0.976747091  0.59339316
#>    47   99 0.392566529 0.976747091 0.615786834 0.916156307  0.56487892
#>    99   14 0.615786834 0.916156307 0.695552975 0.930889044  0.64281904
#>    14    8 0.695552975 0.930889044 0.701136411 0.906191916  0.73322149
#>     8   48 0.701136411 0.906191916 0.854226231 0.994574692  0.80392455
#>    99   70 0.615786834 0.916156307 0.689962513 0.850546515  0.65556900
#>    70   54 0.689962513 0.850546515 0.750517641 0.804054945  0.75091287
#>    54   27 0.750517641 0.804054945 0.821416101 0.801929572  0.78843709
#>    27   48 0.821416101 0.801929572 0.854226231 0.994574692  0.80450184
#>    58   75 0.000000000 0.841474711 0.033519617 0.890690851 -0.69825912
#>    75   77 0.033519617 0.890690851 0.088912749 0.914992771  0.04416036
#>    77   65 0.088912749 0.914992771 0.202760569 0.893106071  0.15550896
#>    65   83 0.202760569 0.893106071 0.273826524 0.996186043  0.17786987
#>    83   16 0.273826524 0.996186043 0.289359929 1.000000000  0.10971050
#>    65   37 0.202760569 0.893106071 0.184314448 0.775095333  0.27603786
#>    37   89 0.184314448 0.775095333 0.340470391 0.701481744  0.29508966
#>    12   47 0.437636470 0.922814087 0.392566529 0.976747091  0.36356936
#>    47   16 0.392566529 0.976747091 0.289359929 1.000000000  0.33492927
#>    34   73 0.520841007 0.384261149 0.530530025 0.432323345  0.49318652
#>    73   39 0.530530025 0.432323345 0.449181897 0.434801744  0.49022076
#>    39   17 0.449181897 0.434801744 0.329690249 0.393045069  0.39217765
#>    55   60 0.437077051 0.741109322 0.534450270 0.821191448  0.44399198
#>    60   23 0.534450270 0.821191448 0.551749264 0.858039997  0.48210478
#>    23   99 0.551749264 0.858039997 0.615786834 0.916156307  0.52780578
#>    34   33 0.520841007 0.384261149 0.668576499 0.359789112  0.59075654
#>    33   71 0.668576499 0.359789112 0.727295976 0.461330952  0.66181309
#>    71   52 0.727295976 0.461330952 0.611341728 0.505508669  0.65692851
#>    52   73 0.611341728 0.505508669 0.530530025 0.432323345  0.56453161
#>    46   62 0.107611584 0.602059228 0.207112675 0.639039478  0.15273077
#>    62   84 0.207112675 0.639039478 0.108107562 0.734442852  0.16064978
#>    84   58 0.108107562 0.734442852 0.000000000 0.841474711 -0.02898535
#>    58   57 0.000000000 0.841474711 0.008205552 0.453984409 -0.13823416
#>    39   45 0.449181897 0.434801744 0.390363427 0.478788419  0.38888688
#>    45   50 0.390363427 0.478788419 0.308579621 0.409562934  0.35113858
#>    74  100 0.494783466 0.017274652 0.984554559 0.012293778  0.73489655
#>   100   32 0.984554559 0.012293778 0.866843740 0.088676758  0.84269452
#>    35   19 0.680527320 0.216714241 0.669923945 0.181024294  0.76728907
#>    19   64 0.669923945 0.181024294 0.646900727 0.096913971  0.71548259
#>    64   74 0.646900727 0.096913971 0.494783466 0.017274652  0.66793684
#>    35   40 0.680527320 0.216714241 0.672840856 0.229177764  0.64986239
#>    40   56 0.672840856 0.229177764 0.507010626 0.359129269  0.52699448
#>    56   26 0.507010626 0.359129269 0.463468399 0.040070187  0.50801955
#>    26   64 0.463468399 0.040070187 0.646900727 0.096913971  0.52175226
#>    42   38 0.189214947 0.136277727 0.196167788 0.004251523  0.22749296
#>    38   41 0.196167788 0.004251523 0.348325591 0.000000000  0.27098421
#>    41   82 0.348325591 0.000000000 0.329580478 0.035738661  0.30570439
#>    82   95 0.329580478 0.035738661 0.386097623 0.072285841  0.32481371
#>    95   24 0.386097623 0.072285841 0.299981229 0.180103353  0.32307461
#>    24   42 0.299981229 0.180103353 0.189214947 0.136277727  0.26285421
#>    90   86 0.974688821 0.049237969 0.934759890 0.477000828  1.50159882
#>    86   28 0.934759890 0.477000828 0.873712427 0.319098038  1.02227838
#>    28   76 0.873712427 0.319098038 0.789548197 0.323367499  0.82769213
#>    76   94 0.789548197 0.323367499 0.750937479 0.303513194  0.81581107
#>    37   36 0.184314448 0.775095333 0.164752244 0.827074168  0.15299512
#>    36   77 0.164752244 0.827074168 0.088912749 0.914992771  0.12642271
#>    84   37 0.108107562 0.734442852 0.184314448 0.775095333  0.13847901
#>    60   70 0.534450270 0.821191448 0.689962513 0.850546515  0.61814989
#>    87   42 0.219389449 0.196444365 0.189214947 0.136277727  0.25054034
#>    95   56 0.386097623 0.072285841 0.507010626 0.359129269  0.45215405
#>    86   53 0.934759890 0.477000828 0.945949709 0.537757937  0.91592534
#>    53   97 0.945949709 0.537757937 0.918105961 0.562637165  0.90344444
#>    97   29 0.918105961 0.562637165 0.841684589 0.581860499  0.86684427
#>    29   71 0.841684589 0.581860499 0.727295976 0.461330952  0.81871646
#>    71   76 0.727295976 0.461330952 0.789548197 0.323367499  0.82274594
#>    95   41 0.386097623 0.072285841 0.348325591 0.000000000  0.40764380
#>    41   74 0.348325591 0.000000000 0.494783466 0.017274652  0.42754857
#>    67   48 0.864522489 0.814954631 0.854226231 0.994574692  0.81463924
#>    54   44 0.750517641 0.804054945 0.804905598 0.780148506  0.78622072
#>    44   67 0.804905598 0.780148506 0.864522489 0.814954631  0.85864590
#>    97   91 0.918105961 0.562637165 0.909859816 0.677547382  0.89152460
#>    91   61 0.909859816 0.677547382 0.822419312 0.625081299  0.88159590
#>    61   49 0.822419312 0.625081299 0.763411248 0.659759006  0.74801100
#>    79   72 0.064542267 0.146066510 0.089909767 0.352175217 -0.07095214
#>    72   69 0.089909767 0.352175217 0.068199061 0.427607207  0.04709614
#>    57   79 0.008205552 0.453984409 0.064542267 0.146066510 -0.60714060
#>    63   98 0.700295871 0.740771609 0.619847134 0.642919415  0.64371052
#>    98   52 0.619847134 0.642919415 0.611341728 0.505508669  0.62332374
#>   100   90 0.984554559 0.012293778 0.974688821 0.049237969  0.89891841
#>    40   96 0.672840856 0.229177764 0.714445487 0.350488494  0.67926889
#>    96   71 0.714445487 0.350488494 0.727295976 0.461330952  0.70224343
#>    94   40 0.750937479 0.303513194 0.672840856 0.229177764  0.72489686
#>    62   89 0.207112675 0.639039478 0.340470391 0.701481744  0.25213097
#>    42   68 0.189214947 0.136277727 0.140763723 0.040364817  0.19997458
#>    73   80 0.530530025 0.432323345 0.369626749 0.575333603  0.49348462
#>    80   45 0.369626749 0.575333603 0.390363427 0.478788419  0.48998789
#>    94   96 0.750937479 0.303513194 0.714445487 0.350488494  0.68754055
#>    59   79 0.090963338 0.171797280 0.064542267 0.146066510  0.12450386
#>    79   68 0.064542267 0.146066510 0.140763723 0.040364817  0.12434120
#>    67   91 0.864522489 0.814954631 0.909859816 0.677547382  0.98584958
#>    91   53 0.909859816 0.677547382 0.945949709 0.537757937  1.22425060
#>    53  100 0.945949709 0.537757937 0.984554559 0.012293778  2.99100778
#>    54   63 0.750517641 0.804054945 0.700295871 0.740771609  0.75776866
#>    49   92 0.763411248 0.659759006 0.842024878 0.675058719  0.79309101
#>    92   67 0.842024878 0.675058719 0.864522489 0.814954631  0.86664694
#>    80   78 0.369626749 0.575333603 0.227973594 0.472072874  0.30767573
#>    78   50 0.227973594 0.472072874 0.308579621 0.409562934  0.30845722
#>    85   88 0.136806574 0.475124385 0.237852316 0.613269142  0.17758090
#>    88   62 0.237852316 0.613269142 0.207112675 0.639039478  0.17592087
#>    61   51 0.822419312 0.625081299 0.819439085 0.630191978  0.76509219
#>    51   92 0.819439085 0.630191978 0.842024878 0.675058719  0.80286508
#>    78   93 0.227973594 0.472072874 0.251334318 0.421545532  0.29151855
#>    61   92 0.822419312 0.625081299 0.842024878 0.675058719  0.84846942
#>    98   80 0.619847134 0.642919415 0.369626749 0.575333603  0.50227236
#>    70   63 0.689962513 0.850546515 0.700295871 0.740771609  0.69601070
#>    98   60 0.619847134 0.642919415 0.534450270 0.821191448  0.54055650
#>    91   92 0.909859816 0.677547382 0.842024878 0.675058719  0.87755722
#>    88   89 0.237852316 0.613269142 0.340470391 0.701481744  0.26865512
#>    89   80 0.340470391 0.701481744 0.369626749 0.575333603  0.34557194
#>    72   85 0.089909767 0.352175217 0.136806574 0.475124385  0.13410576
#>    96   76 0.714445487 0.350488494 0.789548197 0.323367499  0.77465566
#>    93   85 0.251334318 0.421545532 0.136806574 0.475124385  0.16807927
#>    80   88 0.369626749 0.575333603 0.237852316 0.613269142  0.28777113
#>     1   12 0.399756733 0.724013158 0.437636470 0.922814087  0.37279121
#>     1   10 0.399756733 0.724013158 0.289537368 0.922974834  0.36349036
#>     1   89 0.399756733 0.724013158 0.340470391 0.701481744  0.33091589
#>     1   66 0.399756733 0.724013158 0.365108852 0.698697555  0.35786264
#>     1   80 0.399756733 0.724013158 0.369626749 0.575333603  0.43494127
#>     1   55 0.399756733 0.724013158 0.437077051 0.741109322  0.46372399
#>     2   49 0.706708312 0.665344043 0.763411248 0.659759006  0.72537073
#>     2   63 0.706708312 0.665344043 0.700295871 0.740771609  0.73934968
#>     2   31 0.706708312 0.665344043 0.647576029 0.638622528  0.65589100
#>     2   71 0.706708312 0.665344043 0.727295976 0.461330952  0.71719200
#>     3   32 0.869431867 0.121424085 0.866843740 0.088676758  0.89607714
#>     3   90 0.869431867 0.121424085 0.974688821 0.049237969  0.93211633
#>     3   21 0.869431867 0.121424085 0.916110734 0.243279064  0.96866078
#>     3   94 0.869431867 0.121424085 0.750937479 0.303513194  0.81165429
#>     3   35 0.869431867 0.121424085 0.680527320 0.216714241  0.79036619
#>     3   18 0.869431867 0.121424085 0.769697119 0.073718991  0.78031068
#>     3   81 0.869431867 0.121424085 0.863912142 0.107388352  0.79873234
#>     4   72 0.223417079 0.270590358 0.089909767 0.352175217  0.18390234
#>     4   59 0.223417079 0.270590358 0.090963338 0.171797280  0.12660684
#>     4   87 0.223417079 0.270590358 0.219389449 0.196444365  0.14489883
#>     4   24 0.223417079 0.270590358 0.299981229 0.180103353  0.26834205
#>     4   17 0.223417079 0.270590358 0.329690249 0.393045069  0.33133922
#>     4   50 0.223417079 0.270590358 0.308579621 0.409562934  0.26953314
#>     4   93 0.223417079 0.270590358 0.251334318 0.421545532  0.26436721
#>     5   56 0.452886899 0.378645959 0.507010626 0.359129269  0.43218770
#>     5   34 0.452886899 0.378645959 0.520841007 0.384261149  0.48641950
#>     5   11 0.452886899 0.378645959 0.453187123 0.424849812  0.48520426
#>     5   17 0.452886899 0.378645959 0.329690249 0.393045069  0.39319268
#>     5   24 0.452886899 0.378645959 0.299981229 0.180103353  0.37864532
#>     6   30 0.070144705 0.451889512 0.020345608 0.424632565  0.03906091
#>     6   69 0.070144705 0.451889512 0.068199061 0.427607207  0.04329629
#>     6   85 0.070144705 0.451889512 0.136806574 0.475124385  0.11298012
#>     6   46 0.070144705 0.451889512 0.107611584 0.602059228  0.08063757
#>     6   57 0.070144705 0.451889512 0.008205552 0.453984409  0.04207416
#>     7   47 0.717830514 0.961094590 0.392566529 0.976747091  0.59339316
#>     7   99 0.717830514 0.961094590 0.615786834 0.916156307  0.56487892
#>     7   14 0.717830514 0.961094590 0.695552975 0.930889044  0.64281904
#>     7    8 0.717830514 0.961094590 0.701136411 0.906191916  0.73322149
#>     7   48 0.717830514 0.961094590 0.854226231 0.994574692  0.80392455
#>     8   99 0.701136411 0.906191916 0.615786834 0.916156307  0.65826374
#>     8   70 0.701136411 0.906191916 0.689962513 0.850546515  0.65556900
#>     8   54 0.701136411 0.906191916 0.750517641 0.804054945  0.75091287
#>     8   27 0.701136411 0.906191916 0.821416101 0.801929572  0.78843709
#>     9   75 0.081781466 0.977797279 0.033519617 0.890690851 -0.69825912
#>     9   77 0.081781466 0.977797279 0.088912749 0.914992771  0.04416036
#>     9   65 0.081781466 0.977797279 0.202760569 0.893106071  0.15550896
#>     9   83 0.081781466 0.977797279 0.273826524 0.996186043  0.17786987
#>     9   16 0.081781466 0.977797279 0.289359929 1.000000000  0.10971050
#>    10   65 0.289537368 0.922974834 0.202760569 0.893106071  0.23207312
#>    10   37 0.289537368 0.922974834 0.184314448 0.775095333  0.27603786
#>    10   47 0.289537368 0.922974834 0.392566529 0.976747091  0.36356936
#>    10   16 0.289537368 0.922974834 0.289359929 1.000000000  0.33492927
#>    10   83 0.289537368 0.922974834 0.273826524 0.996186043  0.29058045
#>    11   73 0.453187123 0.424849812 0.530530025 0.432323345  0.49318652
#>    11   39 0.453187123 0.424849812 0.449181897 0.434801744  0.49022076
#>    12   60 0.437636470 0.922814087 0.534450270 0.821191448  0.44399198
#>    12   23 0.437636470 0.922814087 0.551749264 0.858039997  0.48210478
#>    12   99 0.437636470 0.922814087 0.615786834 0.916156307  0.52780578
#>    13   33 0.589981855 0.426845358 0.668576499 0.359789112  0.59075654
#>    13   71 0.589981855 0.426845358 0.727295976 0.461330952  0.66181309
#>    13   52 0.589981855 0.426845358 0.611341728 0.505508669  0.65692851
#>    13   73 0.589981855 0.426845358 0.530530025 0.432323345  0.56453161
#>    13   34 0.589981855 0.426845358 0.520841007 0.384261149  0.55769939
#>    15   62 0.102063231 0.653663551 0.207112675 0.639039478  0.15273077
#>    15   84 0.102063231 0.653663551 0.108107562 0.734442852  0.16064978
#>    15   58 0.102063231 0.653663551 0.000000000 0.841474711 -0.02898535
#>    15   57 0.102063231 0.653663551 0.008205552 0.453984409 -0.13823416
#>    15   46 0.102063231 0.653663551 0.107611584 0.602059228 -0.06380307
#>    16   48 0.289359929 1.000000000 0.854226231 0.994574692  0.58198194
#>    17   45 0.329690249 0.393045069 0.390363427 0.478788419  0.38888688
#>    18  100 0.769697119 0.073718991 0.984554559 0.012293778  0.73489655
#>    18   32 0.769697119 0.073718991 0.866843740 0.088676758  0.84269452
#>    18   19 0.769697119 0.073718991 0.669923945 0.181024294  0.76728907
#>    18   64 0.769697119 0.073718991 0.646900727 0.096913971  0.71548259
#>    18   74 0.769697119 0.073718991 0.494783466 0.017274652  0.66793684
#>    19   40 0.669923945 0.181024294 0.672840856 0.229177764  0.64986239
#>    19   56 0.669923945 0.181024294 0.507010626 0.359129269  0.52699448
#>    19   26 0.669923945 0.181024294 0.463468399 0.040070187  0.50801955
#>    20   38 0.295847446 0.041898213 0.196167788 0.004251523  0.22749296
#>    20   41 0.295847446 0.041898213 0.348325591 0.000000000  0.27098421
#>    20   82 0.295847446 0.041898213 0.329580478 0.035738661  0.30570439
#>    20   95 0.295847446 0.041898213 0.386097623 0.072285841  0.32481371
#>    20   24 0.295847446 0.041898213 0.299981229 0.180103353  0.32307461
#>    20   42 0.295847446 0.041898213 0.189214947 0.136277727  0.26285421
#>    21   86 0.916110734 0.243279064 0.934759890 0.477000828  1.50159882
#>    21   28 0.916110734 0.243279064 0.873712427 0.319098038  1.02227838
#>    21   76 0.916110734 0.243279064 0.789548197 0.323367499  0.82769213
#>    22   36 0.122821846 0.812734714 0.164752244 0.827074168  0.15299512
#>    22   77 0.122821846 0.812734714 0.088912749 0.914992771  0.12642271
#>    22   75 0.122821846 0.812734714 0.033519617 0.890690851  0.08181558
#>    22   58 0.122821846 0.812734714 0.000000000 0.841474711  0.06314013
#>    22   84 0.122821846 0.812734714 0.108107562 0.734442852  0.05169288
#>    22   37 0.122821846 0.812734714 0.184314448 0.775095333  0.13847901
#>    23   70 0.551749264 0.858039997 0.689962513 0.850546515  0.61814989
#>    25   53 0.912785856 0.472364770 0.945949709 0.537757937  0.91592534
#>    25   97 0.912785856 0.472364770 0.918105961 0.562637165  0.90344444
#>    25   29 0.912785856 0.472364770 0.841684589 0.581860499  0.86684427
#>    25   71 0.912785856 0.472364770 0.727295976 0.461330952  0.81871646
#>    25   76 0.912785856 0.472364770 0.789548197 0.323367499  0.82274594
#>    25   28 0.912785856 0.472364770 0.873712427 0.319098038  0.83614792
#>    25   86 0.912785856 0.472364770 0.934759890 0.477000828  0.94311189
#>    26   41 0.463468399 0.040070187 0.348325591 0.000000000  0.40764380
#>    26   74 0.463468399 0.040070187 0.494783466 0.017274652  0.42754857
#>    26   95 0.463468399 0.040070187 0.386097623 0.072285841  0.48454067
#>    27   44 0.821416101 0.801929572 0.804905598 0.780148506  0.78622072
#>    27   67 0.821416101 0.801929572 0.864522489 0.814954631  0.85864590
#>    29   91 0.841684589 0.581860499 0.909859816 0.677547382  0.89152460
#>    29   61 0.841684589 0.581860499 0.822419312 0.625081299  0.88159590
#>    29   49 0.841684589 0.581860499 0.763411248 0.659759006  0.74801100
#>    30   72 0.020345608 0.424632565 0.089909767 0.352175217 -0.07095214
#>    30   79 0.020345608 0.424632565 0.064542267 0.146066510 -0.60714060
#>    31   98 0.647576029 0.638622528 0.619847134 0.642919415  0.64371052
#>    31   52 0.647576029 0.638622528 0.611341728 0.505508669  0.62332374
#>    33   56 0.668576499 0.359789112 0.507010626 0.359129269  0.58790987
#>    33   40 0.668576499 0.359789112 0.672840856 0.229177764  0.58806995
#>    33   96 0.668576499 0.359789112 0.714445487 0.350488494  0.67926889
#>    36   65 0.164752244 0.827074168 0.202760569 0.893106071  0.24209350
#>    37   62 0.184314448 0.775095333 0.207112675 0.639039478  0.17363130
#>    38   68 0.196167788 0.004251523 0.140763723 0.040364817  0.19997458
#>    39   80 0.449181897 0.434801744 0.369626749 0.575333603  0.49348462
#>    42   79 0.189214947 0.136277727 0.064542267 0.146066510  0.12450386
#>    42   59 0.189214947 0.136277727 0.090963338 0.171797280  0.15371590
#>    43   67 1.000000000 0.904608439 0.864522489 0.814954631  0.90091726
#>    43   91 1.000000000 0.904608439 0.909859816 0.677547382  0.98584958
#>    43   53 1.000000000 0.904608439 0.945949709 0.537757937  1.22425060
#>    43  100 1.000000000 0.904608439 0.984554559 0.012293778  2.99100778
#>    44   63 0.804905598 0.780148506 0.700295871 0.740771609  0.75776866
#>    44   49 0.804905598 0.780148506 0.763411248 0.659759006  0.76541644
#>    44   92 0.804905598 0.780148506 0.842024878 0.675058719  0.79309101
#>    45   78 0.390363427 0.478788419 0.227973594 0.472072874  0.30767573
#>    46   88 0.107611584 0.602059228 0.237852316 0.613269142  0.17758090
#>    49   51 0.763411248 0.659759006 0.819439085 0.630191978  0.76509219
#>    52   80 0.611341728 0.505508669 0.369626749 0.575333603  0.50227236
#>    53   90 0.945949709 0.537757937 0.974688821 0.049237969  1.82535150
#>    55   98 0.437077051 0.741109322 0.619847134 0.642919415  0.49112204
#>    60   63 0.534450270 0.821191448 0.700295871 0.740771609  0.59864064
#>    78   88 0.227973594 0.472072874 0.237852316 0.613269142  0.28777113
#>    78   85 0.227973594 0.472072874 0.136806574 0.475124385  0.18481469
#>     9   58 0.081781466 0.977797279 0.000000000 0.841474711 -0.69825912
#>    41  100 0.348325591 0.000000000 0.984554559 0.012293778  0.71436033
#>    43   48 1.000000000 0.904608439 0.854226231 0.994574692  0.90091726
#>            my1         mx2           my2 bp1 bp2
#>   0.8321604840  0.44399198  8.319413e-01   0   0
#>   0.8339326776  0.36356936  9.067170e-01   0   0
#>   0.8158873047  0.29508966  8.076490e-01   0   0
#>   0.7449829798  0.34557194  6.362174e-01   0   0
#>   0.6394902893  0.34557194  6.362174e-01   0   0
#>   0.6336574575  0.49112204  6.225098e-01   0   0
#>   0.5641820005  0.74299836  5.609737e-01   0   0
#>   0.7061053935  0.76541644  7.264135e-01   0   0
#>   0.6990101903  0.64371052  7.052966e-01   0   0
#>   0.5633566565  0.69617340  5.539056e-01   0   0
#>   0.1028422851  0.81684904  9.042940e-02   0   0
#>   0.0999939919  0.89891841  9.214517e-03   0   0
#>   0.1532806641  1.50159882  3.141666e-01   0   0
#>   0.2134249894  0.81581107  2.248237e-01   0   0
#>   0.1995717747  0.72489686  2.526796e-01   0   0
#>   0.1796375920  0.76728907  1.715175e-01   0   0
#>   0.1411243306  0.81684904  9.042940e-02   0   0
#>   0.3559571633  0.16807927  3.927774e-01   0   0
#>   0.2621975158 -0.02196966  2.613297e-01   0   0
#>   0.2376731032  0.15371590  1.917309e-01   0   0
#>   0.2309676341  0.25054034  1.431719e-01   0   0
#>   0.2842716933  0.37864532  2.776717e-01   0   0
#>   0.3379105203  0.35113858  4.422062e-01   0   0
#>   0.3410762076  0.29151855  4.707882e-01   0   0
#>   0.2364365324  0.45215405  2.133470e-01   0   0
#>   0.3868322719  0.58790987  3.309809e-01   0   0
#>   0.4015388686  0.49318652  4.148438e-01   0   0
#>   0.4021367432  0.39217765  4.060780e-01   0   0
#>   0.4495598107 -0.60714060  1.822879e-01   0   0
#>   0.4418216675  0.04709614  3.806930e-01   0   0
#>   0.4362381817  0.13410576  4.057360e-01   0   0
#>   0.5290303677  0.17758090  5.513273e-01   0   0
#>   0.5386518185 -0.06380307  6.097296e-01   0   0
#>   1.7626177256  0.58198194  2.058120e+00   0   0
#>   1.1700826076  0.53147832  1.047033e+00   0   0
#>   0.9930999383  0.65826374  9.094791e-01   0   0
#>   0.9264252980  0.65826374  9.094791e-01   0   0
#>   0.9049268168  0.80450184  9.039269e-01   0   0
#>   0.8863975086  0.62044345  8.466860e-01   0   0
#>   0.8672519439  0.69601070  7.957420e-01   0   0
#>   0.8853941742  0.78622072  8.114602e-01   0   0
#>   0.9039268751  0.81463924  9.022003e-01   0   0
#>   1.3530604365  0.06314013  8.344946e-01   0   0
#>   0.9417183804  0.08181558  8.558881e-01   0   0
#>   0.9543617144  0.14200572  8.841220e-01   0   0
#>   0.9863036760  0.23207312  9.489346e-01   0   0
#>   1.6981335115  0.29058045  9.614900e-01   0   0
#>   0.8212051674  0.24209350  8.265110e-01   0   0
#>   0.8076489536  0.25213097  7.165210e-01   0   0
#>   0.9067169672  0.53147832  1.047033e+00   0   0
#>   0.9615921885  0.58198194  2.058120e+00   0   0
#>   0.4148438290  0.55769939  4.018385e-01   0   0
#>   0.4455361930  0.49348462  5.526656e-01   0   0
#>   0.4060780497  0.38888688  4.154950e-01   0   0
#>   0.8319412763  0.54055650  7.145268e-01   0   0
#>   0.8682505517  0.61814989  8.043825e-01   0   0
#>   0.9487622100  0.62044345  8.466860e-01   0   0
#>   0.3481659946  0.58790987  3.309809e-01   0   0
#>   0.4314492755  0.70224343  4.080693e-01   0   0
#>   0.4508986079  0.69617340  5.539056e-01   0   0
#>   0.4759876353  0.49427976  5.535602e-01   0   0
#>   0.6330107500  0.17592087  5.706142e-01   0   0
#>   0.6898955831  0.17363130  7.033672e-01   0   0
#>   0.7040850780  0.05169288  7.855742e-01   0   0
#>   0.6447154118 -2.71778716  5.900904e-01   0   1
#>   0.4154949775  0.48998789  5.506861e-01   0   0
#>   0.4422062005  0.30845722  4.926305e-01   0   0
#>  -0.4544932256  0.71436033 -2.473827e+00   0   0
#>  -0.0774299209  0.89891841  9.214517e-03   0   0
#>   0.1715175001  0.64986239  2.064046e-01   0   0
#>   0.1233474951  0.52175226  1.763766e-01   0   0
#>  -0.1283641600  0.54112909  1.138484e-01   0   0
#>   0.2064046100  0.72489686  2.526796e-01   0   0
#>   0.2138473698  0.58806995  2.917853e-01   0   0
#>   0.1964909202  0.48454067  1.996951e-01   0   0
#>   0.1763766067  0.54112909  1.138484e-01   0   0
#>   0.0720973662  0.19997458  7.064818e-02   0   0
#>  -0.0430573152  0.08749223 -6.610055e+00   0   1
#>   0.0004302439  0.37050834  3.442025e-02   0   0
#>   0.1050831719  0.37050834  3.442025e-02   0   0
#>   0.1102482268  0.45215405  2.133470e-01   0   0
#>   0.1120494486  0.25054034  1.431719e-01   0   0
#>   0.3141666367  1.82535150  3.443869e-01   0   0
#>   0.3524126319  0.94311189  3.830195e-01   0   0
#>   0.2435991377  0.83614792  4.102886e-01   0   0
#>   0.2248236801  0.75454154  3.439747e-01   0   0
#>   0.7929788522  0.24209350  8.265110e-01   0   0
#>   0.8706799820  0.14200572  8.841220e-01   0   0
#>   0.7692634418  0.17363130  7.033672e-01   0   0
#>   0.8043825419  0.62111176  7.886916e-01   0   0
#>   0.1431719002  0.15371590  1.917309e-01   0   0
#>   0.2133470218  0.48454067  1.996951e-01   0   0
#>   0.5118786300  1.82535150  3.443869e-01   0   0
#>   0.5182082610  0.99999615  6.262647e-01   0   0
#>   0.5203652518  0.89152460  6.184806e-01   0   0
#>   0.4891133600  0.74299836  5.609737e-01   0   0
#>   0.4213735992  0.77465566  3.996742e-01   0   0
#>   0.0150156009  0.37050834  3.442025e-02   0   0
#>  -0.0421813077  0.71436033 -2.473827e+00   0   0
#>   0.9022003360  0.90091726  9.071460e-01   0   0
#>   0.8114602187  0.75776866  7.467308e-01   0   0
#>   0.7565604182  0.86664694  7.428560e-01   0   0
#>   0.6184806304  0.99999615  6.262647e-01   0   0
#>   0.6255546606  0.87755722  6.322856e-01   0   0
#>   0.5660104143  0.76509219  5.950760e-01   0   0
#>   0.2673583727 -0.02196966  2.613297e-01   0   0
#>   0.3806930353  0.13410576  4.057360e-01   0   0
#>   0.1822879275 -3.74956856 -3.926513e-01   0   1
#>   0.7052966213  0.59864064  7.423506e-01   0   0
#>   0.5737356159  0.50227236  5.812284e-01   0   0
#>   0.0092145173  2.24690354  3.691863e-01   0   0
#>   0.2947629223  0.68754055  2.919261e-01   0   0
#>   0.4080692740  0.77465566  3.996742e-01   0   0
#>   0.2526796297  0.68754055  2.919261e-01   0   0
#>   0.7165209999  0.26865512  6.812304e-01   0   0
#>   0.0706481773  0.12434120  1.088550e-01   0   0
#>   0.5526655694  0.49427976  5.535602e-01   0   0
#>   0.5506860666  0.30767573  5.115277e-01   0   0
#>   0.2919260802  0.75454154  3.439747e-01   0   0
#>   0.1109266132 -0.02196966  2.613297e-01   0   0
#>   0.1088550229 -6.09979134 -4.379365e+00   0   1
#>   0.7788032157  0.87354139  7.417473e-01   0   0
#>   0.6841612212  0.99999615  6.262647e-01   0   0
#>   0.4238542600  2.24690354  3.691863e-01   0   0
#>   0.7467308138  0.69601070  7.957420e-01   0   0
#>   0.7168749825  0.80286508  6.666535e-01   0   0
#>   0.7428560315  0.87354139  7.417473e-01   0   0
#>   0.5115277336  0.28777113  5.388329e-01   0   0
#>   0.4926304582  0.29151855  4.707882e-01   0   0
#>   0.5513273167  0.18481469  5.460362e-01   0   0
#>   0.5706141856  0.26865512  6.812304e-01   0   0
#>   0.5950759980  0.84846942  6.436964e-01   0   0
#>   0.6666534709  0.84846942  6.436964e-01   0   0
#>   0.4707881916  0.18057920  4.194967e-01   0   0
#>   0.6436963696  0.87755722  6.322856e-01   0   0
#>   0.5812284157  0.49112204  6.225098e-01   0   0
#>   0.7957420405  0.62111176  7.886916e-01   0   0
#>   0.7145268327  0.59864064  7.423506e-01   0   0
#>   0.6322855621  0.87354139  7.417473e-01   0   0
#>   0.6812304243  0.31368471  6.288473e-01   0   0
#>   0.6362173604  0.31368471  6.288473e-01   0   0
#>   0.4057359962  0.16807927  3.927774e-01   0   0
#>   0.3996741818  0.75454154  3.439747e-01   0   0
#>   0.3927773953  0.18057920  4.194967e-01   0   0
#>   0.5388328850  0.31368471  6.288473e-01   0   0
#>   0.8321604840  0.36349036  8.339327e-01   0   0
#>   0.8339326776  0.33091589  8.158873e-01   0   0
#>   0.8158873047  0.35786264  7.449830e-01   0   0
#>   0.7449829798  0.43494127  6.394903e-01   0   0
#>   0.6394902893  0.46372399  6.336575e-01   0   0
#>   0.6336574575  0.37279121  8.321605e-01   0   0
#>   0.5641820005  0.73934968  7.061054e-01   0   0
#>   0.7061053935  0.65589100  6.990102e-01   0   0
#>   0.6990101903  0.71719200  5.633567e-01   0   0
#>   0.5633566565  0.72537073  5.641820e-01   0   0
#>   0.1028422851  0.93211633  9.999399e-02   0   0
#>   0.0999939919  0.96866078  1.532807e-01   0   0
#>   0.1532806641  0.81165429  2.134250e-01   0   0
#>   0.2134249894  0.79036619  1.995718e-01   0   0
#>   0.1995717747  0.78031068  1.796376e-01   0   0
#>   0.1796375920  0.79873234  1.411243e-01   0   0
#>   0.1411243306  0.89607714  1.028423e-01   0   0
#>   0.3559571633  0.12660684  2.621975e-01   0   0
#>   0.2621975158  0.14489883  2.376731e-01   0   0
#>   0.2376731032  0.26834205  2.309676e-01   0   0
#>   0.2309676341  0.33133922  2.842717e-01   0   0
#>   0.2842716933  0.26953314  3.379105e-01   0   0
#>   0.3379105203  0.26436721  3.410762e-01   0   0
#>   0.3410762076  0.18390234  3.559572e-01   0   0
#>   0.2364365324  0.48641950  3.868323e-01   0   0
#>   0.3868322719  0.48520426  4.015389e-01   0   0
#>   0.4015388686  0.39319268  4.021367e-01   0   0
#>   0.4021367432  0.37864532  2.776717e-01   0   0
#>   0.2776716812  0.43218770  2.364365e-01   0   0
#>   0.4495598107  0.04329629  4.418217e-01   0   0
#>   0.4418216675  0.11298012  4.362382e-01   0   0
#>   0.4362381817  0.08063757  5.290304e-01   0   0
#>   0.5290303677  0.04207416  5.386518e-01   0   0
#>   0.5386518185  0.03906091  4.495598e-01   0   0
#>   1.7626177256  0.56487892  1.170083e+00   0   0
#>   1.1700826076  0.64281904  9.930999e-01   0   0
#>   0.9930999383  0.73322149  9.264253e-01   0   0
#>   0.9264252980  0.80392455  9.049268e-01   0   0
#>   0.9049268168  0.59339316  1.762618e+00   0   0
#>   0.9094791247  0.65556900  8.863975e-01   0   0
#>   0.8863975086  0.75091287  8.672519e-01   0   0
#>   0.8672519439  0.78843709  8.853942e-01   0   0
#>   0.8853941742  0.80450184  9.039269e-01   0   0
#>   1.3530604365  0.04416036  9.417184e-01   0   0
#>   0.9417183804  0.15550896  9.543617e-01   0   0
#>   0.9543617144  0.17786987  9.863037e-01   0   0
#>   0.9863036760  0.10971050  1.698134e+00   0   0
#>   1.6981335115 -0.39973914  6.461098e+00   0   1
#>   0.9489345763  0.27603786  8.212052e-01   0   0
#>   0.8212051674  0.29508966  8.076490e-01   0   0
#>   0.9067169672  0.33492927  9.615922e-01   0   0
#>   0.9615921885  0.29058045  9.614900e-01   0   0
#>   0.9614900244  0.23207312  9.489346e-01   0   0
#>   0.4148438290  0.49022076  4.455362e-01   0   0
#>   0.4455361930  0.39217765  4.060780e-01   0   0
#>   0.8319412763  0.48210478  8.682506e-01   0   0
#>   0.8682505517  0.52780578  9.487622e-01   0   0
#>   0.9487622100  0.53147832  1.047033e+00   0   0
#>   0.3481659946  0.66181309  4.314493e-01   0   0
#>   0.4314492755  0.65692851  4.508986e-01   0   0
#>   0.4508986079  0.56453161  4.759876e-01   0   0
#>   0.4759876353  0.55769939  4.018385e-01   0   0
#>   0.4018384634  0.59075654  3.481660e-01   0   0
#>   0.6330107500  0.16064978  6.898956e-01   0   0
#>   0.6898955831 -0.02898535  7.040851e-01   0   0
#>   0.7040850780 -0.13823416  6.447154e-01   0   0
#>   0.6447154118 -0.06380307  6.097296e-01   0   0
#>   0.6097296365  0.15273077  6.330108e-01   0   0
#>   2.0581202247  0.59898367  3.828287e+00   0   1
#>   0.4154949775  0.35113858  4.422062e-01   0   0
#>  -0.4544932256  0.84269452 -7.742992e-02   0   0
#>  -0.0774299209  0.81684904  9.042940e-02   0   0
#>   0.1715175001  0.71548259  1.233475e-01   0   0
#>   0.1233474951  0.66793684 -1.283642e-01   0   0
#>  -0.1283641600  0.73489655 -4.544932e-01   0   0
#>   0.2064046100  0.52699448  2.138474e-01   0   0
#>   0.2138473698  0.50801955  1.964909e-01   0   0
#>   0.1964909202  0.52175226  1.763766e-01   0   0
#>   0.0720973662  0.27098421 -4.305732e-02   0   0
#>  -0.0430573152  0.30570439  4.302439e-04   0   0
#>   0.0004302439  0.32481371  1.050832e-01   0   0
#>   0.1050831719  0.32307461  1.102482e-01   0   0
#>   0.1102482268  0.26285421  1.120494e-01   0   0
#>   0.1120494486  0.22749296  7.209737e-02   0   0
#>   0.3141666367  1.02227838  3.524126e-01   0   0
#>   0.3524126319  0.82769213  2.435991e-01   0   0
#>   0.2435991377  0.81581107  2.248237e-01   0   0
#>   0.7929788522  0.12642271  8.706800e-01   0   0
#>   0.8706799820  0.08181558  8.558881e-01   0   0
#>   0.8558881153  0.06314013  8.344946e-01   0   0
#>   0.8344945557  0.05169288  7.855742e-01   0   0
#>   0.7855741510  0.13847901  7.692634e-01   0   0
#>   0.7692634418  0.15299512  7.929789e-01   0   0
#>   0.8043825419  0.62044345  8.466860e-01   0   0
#>   0.5118786300  0.90344444  5.182083e-01   0   0
#>   0.5182082610  0.86684427  5.203653e-01   0   0
#>   0.5203652518  0.81871646  4.891134e-01   0   0
#>   0.4891133600  0.82274594  4.213736e-01   0   0
#>   0.4213735992  0.83614792  4.102886e-01   0   0
#>   0.4102886447  0.94311189  3.830195e-01   0   0
#>   0.3830195217  0.91592534  5.118786e-01   0   0
#>   0.0150156009  0.42754857 -4.218131e-02   0   0
#>  -0.0421813077  0.54112909  1.138484e-01   0   0
#>   0.1996950995  0.40764380  1.501560e-02   0   0
#>   0.8114602187  0.85864590  7.565604e-01   0   0
#>   0.7565604182  0.81463924  9.022003e-01   0   0
#>   0.6184806304  0.88159590  6.255547e-01   0   0
#>   0.6255546606  0.74801100  5.660104e-01   0   0
#>   0.5660104143  0.74299836  5.609737e-01   0   0
#>   0.2673583727  0.04709614  3.806930e-01   0   0
#>   0.1822879275 -0.07095214  2.673584e-01   0   0
#>   0.7052966213  0.62332374  5.737356e-01   0   0
#>   0.5737356159  0.69617340  5.539056e-01   0   0
#>   0.3309808817  0.58806995  2.917853e-01   0   0
#>   0.2917853490  0.67926889  2.947629e-01   0   0
#>   0.2947629223  0.70224343  4.080693e-01   0   0
#>   0.8265109721  0.14200572  8.841220e-01   0   0
#>   0.7033671859  0.25213097  7.165210e-01   0   0
#>   0.0706481773 -8.05679433 -1.259667e+01   0   1
#>   0.5526655694  0.48998789  5.506861e-01   0   0
#>   0.1109266132  0.12434120  1.088550e-01   0   0
#>   0.1917308568  0.12450386  1.109266e-01   0   0
#>   0.9071460018  0.98584958  7.788032e-01   0   0
#>   0.7788032157  1.22425060  6.841612e-01   0   0
#>   0.6841612212  2.99100778  4.238543e-01   0   0
#>   0.4238542600  4.11135301  4.044617e-01   0   1
#>   0.7467308138  0.76541644  7.264135e-01   0   0
#>   0.7264135085  0.79309101  7.168750e-01   0   0
#>   0.7168749825  0.86664694  7.428560e-01   0   0
#>   0.5115277336  0.30845722  4.926305e-01   0   0
#>   0.5513273167  0.17592087  5.706142e-01   0   0
#>   0.5950759980  0.80286508  6.666535e-01   0   0
#>   0.5812284157  0.49427976  5.535602e-01   0   0
#>   0.3443868812  2.24690354  3.691863e-01   0   0
#>   0.6225098251  0.54055650  7.145268e-01   0   0
#>   0.7423506185  0.62111176  7.886916e-01   0   0
#>   0.5388328850  0.18481469  5.460362e-01   0   0
#>   0.5460361781  0.18057920  4.194967e-01   0   0
#>   1.3530604365 -6.09246197  4.589104e+00   0   1
#>  -2.4738274569  0.74471995 -4.045002e+00   0   1
#>   0.9071460018  3.96684229  5.874914e+00   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.399756733 0.724013158
#>   [2,] 0.706708312 0.665344043
#>   [3,] 0.869431867 0.121424085
#>   [4,] 0.223417079 0.270590358
#>   [5,] 0.452886899 0.378645959
#>   [6,] 0.070144705 0.451889512
#>   [7,] 0.717830514 0.961094590
#>   [8,] 0.701136411 0.906191916
#>   [9,] 0.081781466 0.977797279
#>  [10,] 0.289537368 0.922974834
#>  [11,] 0.453187123 0.424849812
#>  [12,] 0.437636470 0.922814087
#>  [13,] 0.589981855 0.426845358
#>  [14,] 0.695552975 0.930889044
#>  [15,] 0.102063231 0.653663551
#>  [16,] 0.289359929 1.000000000
#>  [17,] 0.329690249 0.393045069
#>  [18,] 0.769697119 0.073718991
#>  [19,] 0.669923945 0.181024294
#>  [20,] 0.295847446 0.041898213
#>  [21,] 0.916110734 0.243279064
#>  [22,] 0.122821846 0.812734714
#>  [23,] 0.551749264 0.858039997
#>  [24,] 0.299981229 0.180103353
#>  [25,] 0.912785856 0.472364770
#>  [26,] 0.463468399 0.040070187
#>  [27,] 0.821416101 0.801929572
#>  [28,] 0.873712427 0.319098038
#>  [29,] 0.841684589 0.581860499
#>  [30,] 0.020345608 0.424632565
#>  [31,] 0.647576029 0.638622528
#>  [32,] 0.866843740 0.088676758
#>  [33,] 0.668576499 0.359789112
#>  [34,] 0.520841007 0.384261149
#>  [35,] 0.680527320 0.216714241
#>  [36,] 0.164752244 0.827074168
#>  [37,] 0.184314448 0.775095333
#>  [38,] 0.196167788 0.004251523
#>  [39,] 0.449181897 0.434801744
#>  [40,] 0.672840856 0.229177764
#>  [41,] 0.348325591 0.000000000
#>  [42,] 0.189214947 0.136277727
#>  [43,] 1.000000000 0.904608439
#>  [44,] 0.804905598 0.780148506
#>  [45,] 0.390363427 0.478788419
#>  [46,] 0.107611584 0.602059228
#>  [47,] 0.392566529 0.976747091
#>  [48,] 0.854226231 0.994574692
#>  [49,] 0.763411248 0.659759006
#>  [50,] 0.308579621 0.409562934
#>  [51,] 0.819439085 0.630191978
#>  [52,] 0.611341728 0.505508669
#>  [53,] 0.945949709 0.537757937
#>  [54,] 0.750517641 0.804054945
#>  [55,] 0.437077051 0.741109322
#>  [56,] 0.507010626 0.359129269
#>  [57,] 0.008205552 0.453984409
#>  [58,] 0.000000000 0.841474711
#>  [59,] 0.090963338 0.171797280
#>  [60,] 0.534450270 0.821191448
#>  [61,] 0.822419312 0.625081299
#>  [62,] 0.207112675 0.639039478
#>  [63,] 0.700295871 0.740771609
#>  [64,] 0.646900727 0.096913971
#>  [65,] 0.202760569 0.893106071
#>  [66,] 0.365108852 0.698697555
#>  [67,] 0.864522489 0.814954631
#>  [68,] 0.140763723 0.040364817
#>  [69,] 0.068199061 0.427607207
#>  [70,] 0.689962513 0.850546515
#>  [71,] 0.727295976 0.461330952
#>  [72,] 0.089909767 0.352175217
#>  [73,] 0.530530025 0.432323345
#>  [74,] 0.494783466 0.017274652
#>  [75,] 0.033519617 0.890690851
#>  [76,] 0.789548197 0.323367499
#>  [77,] 0.088912749 0.914992771
#>  [78,] 0.227973594 0.472072874
#>  [79,] 0.064542267 0.146066510
#>  [80,] 0.369626749 0.575333603
#>  [81,] 0.863912142 0.107388352
#>  [82,] 0.329580478 0.035738661
#>  [83,] 0.273826524 0.996186043
#>  [84,] 0.108107562 0.734442852
#>  [85,] 0.136806574 0.475124385
#>  [86,] 0.934759890 0.477000828
#>  [87,] 0.219389449 0.196444365
#>  [88,] 0.237852316 0.613269142
#>  [89,] 0.340470391 0.701481744
#>  [90,] 0.974688821 0.049237969
#>  [91,] 0.909859816 0.677547382
#>  [92,] 0.842024878 0.675058719
#>  [93,] 0.251334318 0.421545532
#>  [94,] 0.750937479 0.303513194
#>  [95,] 0.386097623 0.072285841
#>  [96,] 0.714445487 0.350488494
#>  [97,] 0.918105961 0.562637165
#>  [98,] 0.619847134 0.642919415
#>  [99,] 0.615786834 0.916156307
#> [100,] 0.984554559 0.012293778
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.3997567,0.7240132) [6]: 10 12 55 66 80 89 
#> 2: (0.7067083,0.665344) [4]: 31 49 63 71 
#> 3: (0.8694319,0.1214241) [7]: 18 21 32 35 81 90 94 
#> 4: (0.2234171,0.2705904) [7]: 17 24 50 59 72 87 93 
#> 5: (0.4528869,0.378646) [5]: 11 17 24 34 56 
#> 6: (0.0701447,0.4518895) [5]: 30 46 57 69 85 
#> 7: (0.7178305,0.9610946) [5]: 8 14 47 48 99 
#> 8: (0.7011364,0.9061919) [7]: 7 14 27 48 54 70 99 
#> 9: (0.08178147,0.9777973) [6]: 16 58 65 75 77 83 
#> 10: (0.2895374,0.9229748) [8]: 1 12 16 37 47 65 83 89 
#> 11: (0.4531871,0.4248498) [5]: 5 17 34 39 73 
#> 12: (0.4376365,0.9228141) [7]: 1 10 23 47 55 60 99 
#> 13: (0.5899819,0.4268454) [5]: 33 34 52 71 73 
#> 14: (0.695553,0.930889) [3]: 7 8 99 
#> 15: (0.1020632,0.6536636) [5]: 46 57 58 62 84 
#> 16: (0.2893599,1) [5]: 9 10 47 48 83 
#> 17: (0.3296902,0.3930451) [7]: 4 5 11 24 39 45 50 
#> 18: (0.7696971,0.07371899) [8]: 3 19 32 35 64 74 81 100 
#> 19: (0.6699239,0.1810243) [6]: 18 26 35 40 56 64 
#> 20: (0.2958474,0.04189821) [6]: 24 38 41 42 82 95 
#> 21: (0.9161107,0.2432791) [6]: 3 28 76 86 90 94 
#> 22: (0.1228218,0.8127347) [6]: 36 37 58 75 77 84 
#> 23: (0.5517493,0.85804) [4]: 12 60 70 99 
#> 24: (0.2999812,0.1801034) [8]: 4 5 17 20 42 56 87 95 
#> 25: (0.9127859,0.4723648) [7]: 28 29 53 71 76 86 97 
#> 26: (0.4634684,0.04007019) [6]: 19 41 56 64 74 95 
#> 27: (0.8214161,0.8019296) [5]: 8 44 48 54 67 
#> 28: (0.8737124,0.319098) [4]: 21 25 76 86 
#> 29: (0.8416846,0.5818605) [6]: 25 49 61 71 91 97 
#> 30: (0.02034561,0.4246326) [5]: 6 57 69 72 79 
#> 31: (0.647576,0.6386225) [5]: 2 52 63 71 98 
#> 32: (0.8668437,0.08867676) [5]: 3 18 81 90 100 
#> 33: (0.6685765,0.3597891) [6]: 13 34 40 56 71 96 
#> 34: (0.520841,0.3842611) [6]: 5 11 13 33 56 73 
#> 35: (0.6805273,0.2167142) [5]: 3 18 19 40 94 
#> 36: (0.1647522,0.8270742) [4]: 22 37 65 77 
#> 37: (0.1843144,0.7750953) [7]: 10 22 36 62 65 84 89 
#> 38: (0.1961678,0.004251523) [4]: 20 41 42 68 
#> 39: (0.4491819,0.4348017) [5]: 11 17 45 73 80 
#> 40: (0.6728409,0.2291778) [6]: 19 33 35 56 94 96 
#> 41: (0.3483256,0) [7]: 20 26 38 74 82 95 100 
#> 42: (0.1892149,0.1362777) [7]: 20 24 38 59 68 79 87 
#> 43: (1,0.9046084) [5]: 48 53 67 91 100 
#> 44: (0.8049056,0.7801485) [6]: 27 49 54 63 67 92 
#> 45: (0.3903634,0.4787884) [5]: 17 39 50 78 80 
#> 46: (0.1076116,0.6020592) [6]: 6 15 57 62 85 88 
#> 47: (0.3925665,0.9767471) [6]: 7 10 12 16 48 99 
#> 48: (0.8542262,0.9945747) [7]: 7 8 16 27 43 47 67 
#> 49: (0.7634112,0.659759) [8]: 2 29 44 51 61 63 71 92 
#> 50: (0.3085796,0.4095629) [5]: 4 17 45 78 93 
#> 51: (0.8194391,0.630192) [3]: 49 61 92 
#> 52: (0.6113417,0.5055087) [6]: 13 31 71 73 80 98 
#> 53: (0.9459497,0.5377579) [7]: 25 43 86 90 91 97 100 
#> 54: (0.7505176,0.8040549) [5]: 8 27 44 63 70 
#> 55: (0.4370771,0.7411093) [5]: 1 12 60 80 98 
#> 56: (0.5070106,0.3591293) [8]: 5 19 24 26 33 34 40 95 
#> 57: (0.008205552,0.4539844) [6]: 6 15 30 46 58 79 
#> 58: (0,0.8414747) [6]: 9 15 22 57 75 84 
#> 59: (0.09096334,0.1717973) [5]: 4 42 72 79 87 
#> 60: (0.5344503,0.8211914) [6]: 12 23 55 63 70 98 
#> 61: (0.8224193,0.6250813) [5]: 29 49 51 91 92 
#> 62: (0.2071127,0.6390395) [6]: 15 37 46 84 88 89 
#> 63: (0.7002959,0.7407716) [8]: 2 31 44 49 54 60 70 98 
#> 64: (0.6469007,0.09691397) [4]: 18 19 26 74 
#> 65: (0.2027606,0.8931061) [6]: 9 10 36 37 77 83 
#> 66: (0.3651089,0.6986976) [3]: 1 80 89 
#> 67: (0.8645225,0.8149546) [6]: 27 43 44 48 91 92 
#> 68: (0.1407637,0.04036482) [3]: 38 42 79 
#> 69: (0.06819906,0.4276072) [4]: 6 30 72 85 
#> 70: (0.6899625,0.8505465) [6]: 8 23 54 60 63 99 
#> 71: (0.727296,0.461331) [10]: 2 13 25 29 31 33 49 52 76 96 
#> 72: (0.08990977,0.3521752) [7]: 4 30 59 69 79 85 93 
#> 73: (0.53053,0.4323233) [6]: 11 13 34 39 52 80 
#> 74: (0.4947835,0.01727465) [5]: 18 26 41 64 100 
#> 75: (0.03351962,0.8906909) [4]: 9 22 58 77 
#> 76: (0.7895482,0.3233675) [6]: 21 25 28 71 94 96 
#> 77: (0.08891275,0.9149928) [5]: 9 22 36 65 75 
#> 78: (0.2279736,0.4720729) [6]: 45 50 80 85 88 93 
#> 79: (0.06454227,0.1460665) [6]: 30 42 57 59 68 72 
#> 80: (0.3696267,0.5753336) [11]: 1 39 45 52 55 66 73 78 88 89 98 
#> 81: (0.8639121,0.1073884) [3]: 3 18 32 
#> 82: (0.3295805,0.03573866) [3]: 20 41 95 
#> 83: (0.2738265,0.996186) [4]: 9 10 16 65 
#> 84: (0.1081076,0.7344429) [5]: 15 22 37 58 62 
#> 85: (0.1368066,0.4751244) [7]: 6 46 69 72 78 88 93 
#> 86: (0.9347599,0.4770008) [5]: 21 25 28 53 90 
#> 87: (0.2193894,0.1964444) [4]: 4 24 42 59 
#> 88: (0.2378523,0.6132691) [6]: 46 62 78 80 85 89 
#> 89: (0.3404704,0.7014817) [7]: 1 10 37 62 66 80 88 
#> 90: (0.9746888,0.04923797) [6]: 3 21 32 53 86 100 
#> 91: (0.9098598,0.6775474) [7]: 29 43 53 61 67 92 97 
#> 92: (0.8420249,0.6750587) [6]: 44 49 51 61 67 91 
#> 93: (0.2513343,0.4215455) [5]: 4 50 72 78 85 
#> 94: (0.7509375,0.3035132) [6]: 3 21 35 40 76 96 
#> 95: (0.3860976,0.07228584) [6]: 20 24 26 41 56 82 
#> 96: (0.7144455,0.3504885) [5]: 33 40 71 76 94 
#> 97: (0.918106,0.5626372) [4]: 25 29 53 91 
#> 98: (0.6198471,0.6429194) [6]: 31 52 55 60 63 80 
#> 99: (0.6157868,0.9161563) [7]: 7 8 12 14 23 47 70 
#> 100: (0.9845546,0.01229378) [7]: 18 32 41 43 53 74 90 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  9 16 38 41 43 48 57 58 68 79 100 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.181705626 0.148099190 0.227273760 0.024795271 0.123446653 0.178972434 
#>                                                                         
#> 0.201687891 0.102696606 0.114951366 0.194390260 0.018939853 0.114830220 
#>                                                                         
#> 0.202690253 0.175813409 0.111766045 0.168519716 0.100050469 0.175698965 
#>                                                                         
#> 0.180381013 0.130769815 0.082231769 0.215004187 0.026804822 0.058485958 
#>                                                                         
#> 0.273699561 0.028686073 0.078895421 0.127526545 0.031763370 0.047945818 
#>                                                                         
#> 0.083455815 0.130248999 0.178347165 0.462003793 0.231297531 0.081115293 
#>                                                                         
#> 0.025320404 0.176771062 0.099028663 0.076343891 0.070930310 0.195419157 
#>                                                                         
#> 0.059546563 0.060489523 0.115932539 0.125203238 0.015994778 0.119443684 
#>                                                                         
#> 0.172637305 0.070285621 0.105793667 0.049029091 0.081385873 0.126577542 
#>                                                                         
#> 0.126074147 0.040707134 0.086477257 0.149748643 0.117297580 0.124084884 
#>                                                                         
#> 0.109025790 0.106150865 0.137491150 0.152128445 0.387577174 0.073446851 
#>                                                                         
#> 0.107148303 0.489796419 0.140321760 0.037231760 0.087204443 0.171703472 
#>                                                                         
#> 0.014643126 0.210682364 0.322016495 0.192038108 0.132209155 0.152217189 
#>                                                                         
#> 0.040356303 0.067304413 0.137987859 0.119121177 0.429622373 0.169292893 
#>                                                                         
#> 0.084272450 0.043416368 0.055538087 0.116109042 0.086371949 0.158258578 
#>                                                                         
#> 0.067309174 0.311286214 0.061778947 0.037339662 0.078802047 0.166168970 
#>                                                                         
#> 0.151358030 0.081559605 0.147473125 0.179914922 0.059410165 0.069033616 
#>                                                                         
#> 0.115205715 0.101973191 0.068443370 0.207663933 0.078494204 0.313029164 
#>                                                                         
#> 0.126676955 0.137673727 0.038238802 0.128246788 0.111584880 0.107818545 
#>                                                                         
#> 0.147252561 0.107456072 0.215271452 0.098747063 0.059483984 0.036880151 
#>                                                                         
#> 0.130317144 0.144693557 0.144373022 0.526880363 0.080789893 0.080088600 
#>                                                                         
#> 0.141693362 0.175295164 0.102004040 0.171155531 0.040112787 0.005916147 
#>                                                                         
#> 0.050230891 0.055666289 0.053685386 0.259187351 0.110260184 0.197670286 
#>                                                                         
#> 0.067880574 0.135321589 0.129473730 0.131589546 0.079849643 0.126440886 
#>                                                                         
#> 0.137126242 0.202377577 0.227451219 0.063423458 0.042911017 0.151701766 
#>                                                                         
#> 0.041049786 0.056977325 0.075699651 0.064889647 0.205049247 0.032849442 
#>                                                                         
#> 0.127631742 0.130489665 0.217249543 0.211577744 0.110556754 0.015082081 
#>                                                                         
#> 0.156461789 0.165239419 0.074255304 0.118532557 0.162139270 0.162990906 
#>                                                                         
#> 0.153514940 0.057535024 0.068185710 0.046204828 0.124035273 0.250597906 
#>                                                                         
#> 0.056770513 0.024360128 0.070595072 0.154773094 0.061974569 0.325640386 
#>                                                                         
#> 0.111500502 0.037532170 0.057384638 0.140444683 0.085929270 0.056756204 
#>                                                                         
#> 0.113448079 0.159178642 0.099582809 0.063208081 0.147677162 0.192923433 
#>                                                                         
#> 0.208762494 0.091773394 0.181494380 0.116217312 0.077025370 0.074877979 
#>                                                                         
#> 0.077703141 0.010727665 0.140356947 0.131215138 0.178274727 0.103313397 
#>                                                                         
#> 0.141578331 0.081511721 0.059703672 0.081202659 0.106062478 0.081005120 
#>                                                                         
#> 0.213752041 0.220637766 0.051901737 0.564892356 0.105038834 0.223465380 
#>                                                                         
#> 0.098291407 0.146523425 0.124967840 0.280648321 0.048241736 0.241375500 
#>                                                                         
#> 0.249983905 0.106551900 0.067152185 0.034290779 0.095228685 0.138266948 
#>                                                                         
#> 0.142400781 0.234464612 0.086868483 0.149773941 0.044314538 0.107733639 
#>                                                                         
#> 0.118541332 0.126139578 0.079662575 0.072097594 0.138416236 0.073321944 
#>                                                                         
#> 0.090429027 0.130555371 0.185817762 0.193359011 0.158168972 0.022457765 
#>                                                                         
#> 0.121915897 0.038733317 0.083809816 0.027331513 0.045031243 0.117489750 
#>                                                                         
#> 0.047320064 0.110430490 0.100445206 0.282050335 0.028059844 0.137957327 
#>                                                                         
#> 0.161567220 0.130680944 0.046802410 0.076189533 0.137952727 0.066134564 
#>                                                                         
#> 0.161487538 0.125056377 0.104474960 0.162456028 0.244298948 0.370810902 
#>                                                                         
#> 0.892448327 0.111775377 0.127339753 0.111452700 0.162528633 0.130722265 
#>                                                                         
#> 0.063350829 0.251598196 0.489364584 0.207475688 0.184315257 0.141541426 
#>                                                 
#> 0.091218075 0.158971855 0.636347732 0.171300667 
#> 
#> $alpha
#> [1] 0.2959123
#> 
#> attr(,"class")
#> [1] "scree"