Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2         mx1
#>    87   55 0.332724529 0.792817060 0.327012023 0.822797390  0.31057312
#>     1   55 0.296339956 0.783732159 0.327012023 0.822797390  0.29040272
#>     1   31 0.296339956 0.783732159 0.280035323 0.855192000  0.26729968
#>    87    1 0.332724529 0.792817060 0.296339956 0.783732159  0.33097586
#>    85   35 0.204097553 0.739421867 0.262968802 0.704142502  0.25340539
#>    59   85 0.234462308 0.841599997 0.204097553 0.739421867  0.23915740
#>    97   67 0.185735192 0.739555165 0.185004849 0.716574391  0.19483126
#>    53   97 0.163978197 0.805819642 0.185735192 0.739555165  0.19520587
#>    59   53 0.234462308 0.841599997 0.163978197 0.805819642  0.21550355
#>    87   73 0.332724529 0.792817060 0.415229328 0.699972898  0.43904800
#>    87   35 0.332724529 0.792817060 0.262968802 0.704142502  0.33947359
#>    59   31 0.234462308 0.841599997 0.280035323 0.855192000  0.22283224
#>    31   55 0.280035323 0.855192000 0.327012023 0.822797390  0.33582490
#>    52   72 0.476392917 0.904419636 0.350568913 0.947756456  0.39572766
#>    67   35 0.185004849 0.716574391 0.262968802 0.704142502  0.22300772
#>    29   67 0.179238057 0.695582525 0.185004849 0.716574391  0.14844700
#>    29   32 0.179238057 0.695582525 0.171978083 0.687323736  0.21371824
#>     3   73 0.215795275 0.606955652 0.415229328 0.699972898  0.33653407
#>    35   29 0.262968802 0.704142502 0.179238057 0.695582525  0.22491584
#>    94   12 0.418434738 1.000000000 0.100390918 0.955169841  0.23201390
#>    77   94 0.438222683 0.988184714 0.418434738 1.000000000  0.40238668
#>    52   77 0.476392917 0.904419636 0.438222683 0.988184714  0.41358019
#>    52   73 0.476392917 0.904419636 0.415229328 0.699972898  0.46304222
#>    95   73 0.529225247 0.882645942 0.415229328 0.699972898  0.47577095
#>     6    7 0.513027197 0.613193178 0.536804302 0.700308919  0.47609991
#>    12   50 0.100390918 0.955169841 0.045728987 0.825450735  0.11436144
#>    31   72 0.280035323 0.855192000 0.350568913 0.947756456  0.22601471
#>    50   93 0.045728987 0.825450735 0.012199727 0.812712780  0.03865709
#>    53   50 0.163978197 0.805819642 0.045728987 0.825450735  0.09353312
#>     3    5 0.215795275 0.606955652 0.487077318 0.564640074  0.35034260
#>     5    6 0.487077318 0.564640074 0.513027197 0.613193178  0.43495682
#>     4   69 0.004017651 0.710111042 0.039924261 0.673309648  0.07963842
#>    69   32 0.039924261 0.673309648 0.171978083 0.687323736  0.10023854
#>    93   12 0.012199727 0.812712780 0.100390918 0.955169841 -3.47658075
#>     3   32 0.215795275 0.606955652 0.171978083 0.687323736  0.11408910
#>     3   78 0.215795275 0.606955652 0.186767409 0.516800182  0.28752344
#>     3   69 0.215795275 0.606955652 0.039924261 0.673309648  0.10948625
#>    18    4 0.010371495 0.809890943 0.004017651 0.710111042  0.05476203
#>     3   38 0.215795275 0.606955652 0.324250271 0.410812809  0.34599126
#>    38    2 0.324250271 0.410812809 0.484100448 0.518068191  0.34660756
#>    78   69 0.186767409 0.516800182 0.039924261 0.673309648  0.07909967
#>    83   38 0.248792157 0.422524602 0.324250271 0.410812809  0.28577077
#>    83   25 0.248792157 0.422524602 0.256410586 0.436727306  0.19989683
#>     3   25 0.215795275 0.606955652 0.256410586 0.436727306  0.33698818
#>    77   95 0.438222683 0.988184714 0.529225247 0.882645942  0.58664456
#>    95    7 0.529225247 0.882645942 0.536804302 0.700308919  0.64189965
#>    68    7 0.782573553 0.775673342 0.536804302 0.700308919  0.65202011
#>    78    9 0.186767409 0.516800182 0.033408510 0.455071852  0.13537761
#>    83   78 0.248792157 0.422524602 0.186767409 0.516800182  0.14644256
#>    20   84 0.281685325 0.342672148 0.285330717 0.256195776  0.34500672
#>    20   38 0.281685325 0.342672148 0.324250271 0.410812809  0.28232563
#>     6   96 0.513027197 0.613193178 0.695496899 0.409466047  0.65648456
#>    90   47 0.813417588 0.596557867 0.762007821 0.691532787  0.66161150
#>    34   95 0.720784139 0.951154354 0.529225247 0.882645942  0.65616171
#>    20   83 0.281685325 0.342672148 0.248792157 0.422524602  0.22576919
#>    71    9 0.180271346 0.326131407 0.033408510 0.455071852  0.11648358
#>    43   49 0.013685040 0.339178137 0.029817269 0.354139084 -0.03415874
#>    34   68 0.720784139 0.951154354 0.782573553 0.775673342  0.76703230
#>    15   68 0.841522451 0.787936158 0.782573553 0.775673342  0.81982210
#>    84   71 0.285330717 0.256195776 0.180271346 0.326131407  0.22412155
#>    71   49 0.180271346 0.326131407 0.029817269 0.354139084  0.11637507
#>    13   47 0.858055326 0.774460188 0.762007821 0.691532787  0.84990764
#>    54   64 0.323044573 0.230058808 0.360587265 0.213945642  0.38199195
#>    13   75 0.858055326 0.774460188 0.861397765 0.865215218  0.88065574
#>    68   47 0.782573553 0.775673342 0.762007821 0.691532787  0.81946272
#>     6    2 0.513027197 0.613193178 0.484100448 0.518068191  0.62273181
#>    51   43 0.130932456 0.303655266 0.013685040 0.339178137  0.05741118
#>    51   49 0.130932456 0.303655266 0.029817269 0.354139084  0.06562356
#>     4   43 0.004017651 0.710111042 0.013685040 0.339178137 -0.62819209
#>    96   91 0.695496899 0.409466047 0.721106931 0.416219379  0.67654985
#>    96   33 0.695496899 0.409466047 0.610827860 0.316535263  0.68926457
#>    96    2 0.695496899 0.409466047 0.484100448 0.518068191  0.57473861
#>    28   89 0.823766229 0.548431046 0.856795203 0.599054909  0.83634375
#>     8   28 0.793567461 0.488653736 0.823766229 0.548431046  0.73671917
#>    64   38 0.360587265 0.213945642 0.324250271 0.410812809  0.39150153
#>    71   51 0.180271346 0.326131407 0.130932456 0.303655266  0.17991403
#>    90    6 0.813417588 0.596557867 0.513027197 0.613193178  0.66122333
#>    99   13 0.930833492 0.814898842 0.858055326 0.774460188  0.91619746
#>    17   89 0.860875565 0.547857460 0.856795203 0.599054909  0.84269225
#>    39   64 0.357432673 0.214384919 0.360587265 0.213945642  0.36888304
#>    39   54 0.357432673 0.214384919 0.323044573 0.230058808  0.31361450
#>    64   44 0.360587265 0.213945642 0.468433844 0.197546142  0.43327567
#>    39  100 0.357432673 0.214384919 0.376790209 0.141746597  0.35347715
#>    54   84 0.323044573 0.230058808 0.285330717 0.256195776  0.27608607
#>    99   75 0.930833492 0.814898842 0.861397765 0.865215218  0.88089871
#>     8   37 0.793567461 0.488653736 0.838375337 0.500479961  0.80979375
#>    37   17 0.838375337 0.500479961 0.860875565 0.547857460  0.84200624
#>    91   42 0.721106931 0.416219379 0.857347999 0.455088034  0.79526344
#>    42   37 0.857347999 0.455088034 0.838375337 0.500479961  0.82312932
#>    13   89 0.858055326 0.774460188 0.856795203 0.599054909  0.86819510
#>    98   36 0.472788951 0.160676808 0.582485723 0.182081028  0.52487780
#>    38   44 0.324250271 0.410812809 0.468433844 0.197546142  0.46273283
#>    98   44 0.472788951 0.160676808 0.468433844 0.197546142  0.42045622
#>    48   42 0.860571939 0.375255213 0.857347999 0.455088034  0.79579856
#>    22   70 0.965742888 0.620471102 0.954065857 0.707521436  0.90213162
#>    36   33 0.582485723 0.182081028 0.610827860 0.316535263  0.53528138
#>    51   45 0.130932456 0.303655266 0.000000000 0.216604650  0.08010473
#>    16   51 0.222709033 0.213121496 0.130932456 0.303655266  0.14369758
#>    60  100 0.238489842 0.128945044 0.376790209 0.141746597  0.30546447
#>    46   36 0.607083212 0.102881328 0.582485723 0.182081028  0.53679036
#>    16   60 0.222709033 0.213121496 0.238489842 0.128945044  0.27985245
#>    34   75 0.720784139 0.951154354 0.861397765 0.865215218  0.87649979
#>    99   70 0.930833492 0.814898842 0.954065857 0.707521436  1.06711732
#>    22   10 0.965742888 0.620471102 1.000000000 0.594264934  1.03377109
#>    17   22 0.860875565 0.547857460 0.965742888 0.620471102  0.91748712
#>    40   62 0.641735341 0.204503902 0.690607266 0.205501831  0.66482059
#>    17   66 0.860875565 0.547857460 0.917209716 0.471998148  0.89490014
#>    57   48 0.768855605 0.328764965 0.860571939 0.375255213  0.79146656
#>    66   42 0.917209716 0.471998148 0.857347999 0.455088034  0.89524253
#>    56   57 0.762522784 0.255334067 0.768855605 0.328764965  0.69171619
#>    40   36 0.641735341 0.204503902 0.582485723 0.182081028  0.61608680
#>    62   56 0.690607266 0.205501831 0.762522784 0.255334067  0.68719810
#>    30   66 0.942582717 0.510865056 0.917209716 0.471998148  1.06350529
#>    48   74 0.860571939 0.375255213 0.793632704 0.278676037  0.84187788
#>    10   88 1.000000000 0.594264934 0.983195595 0.898836683  1.24773325
#>    22   30 0.965742888 0.620471102 0.942582717 0.510865056  0.95141568
#>    10   30 1.000000000 0.594264934 0.942582717 0.510865056  1.16168339
#>    86   79 0.100401300 0.157692140 0.056788456 0.125604826  0.12247579
#>    86   45 0.100401300 0.157692140 0.000000000 0.216604650  0.04794610
#>    81   92 0.436101951 0.090614382 0.515737018 0.029036693  0.50595757
#>   100   81 0.376790209 0.141746597 0.436101951 0.090614382  0.42699677
#>    60   58 0.238489842 0.128945044 0.206528194 0.082423443  0.27249762
#>    56   65 0.762522784 0.255334067 0.718778761 0.205308164  0.75299978
#>    27   65 0.734058789 0.165865257 0.718778761 0.205308164  0.70449839
#>    60   86 0.238489842 0.128945044 0.100401300 0.157692140  0.16923097
#>    46   82 0.607083212 0.102881328 0.656035908 0.121674086  0.62697936
#>    41   36 0.631651753 0.152998736 0.582485723 0.182081028  0.59167497
#>    82   27 0.656035908 0.121674086 0.734058789 0.165865257  0.68702293
#>    41   82 0.631651753 0.152998736 0.656035908 0.121674086  0.67609700
#>    19   74 0.861753605 0.258584744 0.793632704 0.278676037  0.81334908
#>    11   19 0.758941593 0.182946456 0.861753605 0.258584744  0.81346256
#>    19   26 0.861753605 0.258584744 0.881236228 0.173027045  0.92508041
#>    61   48 0.899323911 0.392821035 0.860571939 0.375255213  0.91014692
#>   100   80 0.376790209 0.141746597 0.298606407 0.009752332  0.34934443
#>    21   80 0.171531035 0.020286502 0.298606407 0.009752332  0.23589280
#>    58   86 0.206528194 0.082423443 0.100401300 0.157692140  0.12879834
#>    23   82 0.716380395 0.104371984 0.656035908 0.121674086  0.66272087
#>    92   23 0.515737018 0.029036693 0.716380395 0.104371984  0.66397592
#>    92   46 0.515737018 0.029036693 0.607083212 0.102881328  0.52891087
#>    27   23 0.734058789 0.165865257 0.716380395 0.104371984  0.77029660
#>    11   27 0.758941593 0.182946456 0.734058789 0.165865257  0.77550429
#>    24   80 0.393375525 0.004599453 0.298606407 0.009752332  0.28112325
#>    21   79 0.171531035 0.020286502 0.056788456 0.125604826  0.11840289
#>    79   45 0.056788456 0.125604826 0.000000000 0.216604650 -2.40753602
#>    66   61 0.917209716 0.471998148 0.899323911 0.392821035  1.05114904
#>    26   11 0.881236228 0.173027045 0.758941593 0.182946456  0.82012889
#>    61   19 0.899323911 0.392821035 0.861753605 0.258584744  0.92225300
#>    26   76 0.881236228 0.173027045 0.826879867 0.117502109  1.71585740
#>    35    1 0.262968802 0.704142502 0.296339956 0.783732159  0.33097586
#>     1   85 0.296339956 0.783732159 0.204097553 0.739421867  0.23915740
#>    85   67 0.204097553 0.739421867 0.185004849 0.716574391  0.19483126
#>    85   97 0.204097553 0.739421867 0.185735192 0.739555165  0.19520587
#>    85   53 0.204097553 0.739421867 0.163978197 0.805819642  0.21550355
#>    73   35 0.415229328 0.699972898 0.262968802 0.704142502  0.33947359
#>    12   59 0.100390918 0.955169841 0.234462308 0.841599997  0.16360350
#>    12   31 0.100390918 0.955169841 0.280035323 0.855192000  0.22283224
#>    72   55 0.350568913 0.947756456 0.327012023 0.822797390  0.33582490
#>    32   67 0.171978083 0.687323736 0.185004849 0.716574391  0.14844700
#>    69   53 0.039924261 0.673309648 0.163978197 0.805819642  0.09468738
#>    32   97 0.171978083 0.687323736 0.185735192 0.739555165  0.11481570
#>    52   87 0.476392917 0.904419636 0.332724529 0.792817060  0.42254379
#>     3   29 0.215795275 0.606955652 0.179238057 0.695582525  0.22491584
#>    72   12 0.350568913 0.947756456 0.100390918 0.955169841  0.23201390
#>    72   94 0.350568913 0.947756456 0.418434738 1.000000000  0.40238668
#>    72   77 0.350568913 0.947756456 0.438222683 0.988184714  0.41358019
#>    95   52 0.529225247 0.882645942 0.476392917 0.904419636  0.54010635
#>     7   73 0.536804302 0.700308919 0.415229328 0.699972898  0.47577095
#>    18   93 0.010371495 0.809890943 0.012199727 0.812712780  0.03865709
#>    69   50 0.039924261 0.673309648 0.045728987 0.825450735  0.09353312
#>    73    5 0.415229328 0.699972898 0.487077318 0.564640074  0.35034260
#>    73    6 0.415229328 0.699972898 0.513027197 0.613193178  0.43495682
#>    18   12 0.010371495 0.809890943 0.100390918 0.955169841 -3.47658075
#>    25   78 0.256410586 0.436727306 0.186767409 0.516800182  0.28752344
#>     9    4 0.033408510 0.455071852 0.004017651 0.710111042 -0.10440945
#>    50    4 0.045728987 0.825450735 0.004017651 0.710111042  0.05476203
#>     2    5 0.484100448 0.518068191 0.487077318 0.564640074  0.60309681
#>     5   38 0.487077318 0.564640074 0.324250271 0.410812809  0.34599126
#>     9   69 0.033408510 0.455071852 0.039924261 0.673309648  0.07909967
#>    25   38 0.256410586 0.436727306 0.324250271 0.410812809  0.28577077
#>    34   77 0.720784139 0.951154354 0.438222683 0.988184714  0.62611483
#>    47    6 0.762007821 0.691532787 0.513027197 0.613193178  0.64658551
#>    47    7 0.762007821 0.691532787 0.536804302 0.700308919  0.65202011
#>    71   78 0.180271346 0.326131407 0.186767409 0.516800182  0.14644256
#>     8   96 0.793567461 0.488653736 0.695496899 0.409466047  0.65648456
#>    68   95 0.782573553 0.775673342 0.529225247 0.882645942  0.65616171
#>    71   83 0.180271346 0.326131407 0.248792157 0.422524602  0.22576919
#>    49    9 0.029817269 0.354139084 0.033408510 0.455071852  0.11648358
#>    75   68 0.861397765 0.865215218 0.782573553 0.775673342  0.76703230
#>    13   68 0.858055326 0.774460188 0.782573553 0.775673342  0.81982210
#>    16   71 0.222709033 0.213121496 0.180271346 0.326131407  0.22412155
#>    89   47 0.856795203 0.599054909 0.762007821 0.691532787  0.84990764
#>    20   64 0.281685325 0.342672148 0.360587265 0.213945642  0.38199195
#>    15   75 0.841522451 0.787936158 0.861397765 0.865215218  0.88065574
#>    71   20 0.180271346 0.326131407 0.281685325 0.342672148  0.23700138
#>    33   38 0.610827860 0.316535263 0.324250271 0.410812809  0.46893663
#>    45   43 0.000000000 0.216604650 0.013685040 0.339178137  0.05741118
#>     9   43 0.033408510 0.455071852 0.013685040 0.339178137 -0.62819209
#>     8   91 0.793567461 0.488653736 0.721106931 0.416219379  0.67654985
#>    57   33 0.768855605 0.328764965 0.610827860 0.316535263  0.68926457
#>    33    2 0.610827860 0.316535263 0.484100448 0.518068191  0.57473861
#>    90   89 0.813417588 0.596557867 0.856795203 0.599054909  0.83634375
#>    90   28 0.813417588 0.596557867 0.823766229 0.548431046  0.73671917
#>     8    6 0.793567461 0.488653736 0.513027197 0.613193178  0.66122333
#>    70   13 0.954065857 0.707521436 0.858055326 0.774460188  0.91619746
#>   100   54 0.376790209 0.141746597 0.323044573 0.230058808  0.31361450
#>    64  100 0.360587265 0.213945642 0.376790209 0.141746597  0.35347715
#>    16   84 0.222709033 0.213121496 0.285330717 0.256195776  0.27608607
#>    28   37 0.823766229 0.548431046 0.838375337 0.500479961  0.80979375
#>    28   17 0.823766229 0.548431046 0.860875565 0.547857460  0.84200624
#>     8   42 0.793567461 0.488653736 0.857347999 0.455088034  0.79526344
#>    70   89 0.954065857 0.707521436 0.856795203 0.599054909  0.86819510
#>    44   36 0.468433844 0.197546142 0.582485723 0.182081028  0.52487780
#>    33   44 0.610827860 0.316535263 0.468433844 0.197546142  0.46273283
#>   100   44 0.376790209 0.141746597 0.468433844 0.197546142  0.42045622
#>    66   37 0.917209716 0.471998148 0.838375337 0.500479961  0.88536399
#>    57   96 0.768855605 0.328764965 0.695496899 0.409466047  0.72221944
#>    86   51 0.100401300 0.157692140 0.130932456 0.303655266  0.14369758
#>    45    4 0.000000000 0.216604650 0.004017651 0.710111042 -1.78460944
#>    54   60 0.323044573 0.230058808 0.238489842 0.128945044  0.27985245
#>    88   75 0.983195595 0.898836683 0.861397765 0.865215218  0.87649979
#>    88   99 0.983195595 0.898836683 0.930833492 0.814898842  0.92346398
#>    88   70 0.983195595 0.898836683 0.954065857 0.707521436  1.06711732
#>    86   16 0.100401300 0.157692140 0.222709033 0.213121496  0.17296586
#>    70   10 0.954065857 0.707521436 1.000000000 0.594264934  1.03377109
#>    89   22 0.856795203 0.599054909 0.965742888 0.620471102  0.91748712
#>    33   62 0.610827860 0.316535263 0.690607266 0.205501831  0.66482059
#>    91   48 0.721106931 0.416219379 0.860571939 0.375255213  0.79146656
#>    30   17 0.942582717 0.510865056 0.860875565 0.547857460  0.92130976
#>    61   42 0.899323911 0.392821035 0.857347999 0.455088034  0.89524253
#>    74   56 0.793632704 0.278676037 0.762522784 0.255334067  0.75884377
#>    33   56 0.610827860 0.316535263 0.762522784 0.255334067  0.68719810
#>    74   57 0.793632704 0.278676037 0.768855605 0.328764965  0.82757449
#>    14   66 0.986262629 0.262677277 0.917209716 0.471998148  1.06350529
#>    14   30 0.986262629 0.262677277 0.942582717 0.510865056  1.16168339
#>    98   92 0.472788951 0.160676808 0.515737018 0.029036693  0.50595757
#>    98   81 0.472788951 0.160676808 0.436101951 0.090614382  0.42699677
#>    80   60 0.298606407 0.009752332 0.238489842 0.128945044  0.31173284
#>    80   58 0.298606407 0.009752332 0.206528194 0.082423443  0.27249762
#>    11   65 0.758941593 0.182946456 0.718778761 0.205308164  0.75299978
#>    65   62 0.718778761 0.205308164 0.690607266 0.205501831  0.70507814
#>    41   40 0.631651753 0.152998736 0.641735341 0.204503902  0.66682781
#>    62   27 0.690607266 0.205501831 0.734058789 0.165865257  0.68702293
#>    62   82 0.690607266 0.205501831 0.656035908 0.121674086  0.67609700
#>    56   19 0.762522784 0.255334067 0.861753605 0.258584744  0.81346256
#>    14   26 0.986262629 0.262677277 0.881236228 0.173027045  0.92508041
#>    19   48 0.861753605 0.258584744 0.860571939 0.375255213  0.91014692
#>    24  100 0.393375525 0.004599453 0.376790209 0.141746597  0.36754084
#>    21   86 0.171531035 0.020286502 0.100401300 0.157692140  0.12879834
#>    46   23 0.607083212 0.102881328 0.716380395 0.104371984  0.66397592
#>    98   46 0.472788951 0.160676808 0.607083212 0.102881328  0.52891087
#>    24   81 0.393375525 0.004599453 0.436101951 0.090614382  0.45211523
#>    76   23 0.826879867 0.117502109 0.716380395 0.104371984  0.77029660
#>    76   27 0.826879867 0.117502109 0.734058789 0.165865257  0.77550429
#>    63   80 0.127727882 0.000000000 0.298606407 0.009752332  0.28112325
#>    63   21 0.127727882 0.000000000 0.171531035 0.020286502  0.22175458
#>    63   79 0.127727882 0.000000000 0.056788456 0.125604826  0.11840289
#>    63   45 0.127727882 0.000000000 0.000000000 0.216604650 -2.40753602
#>    14   61 0.986262629 0.262677277 0.899323911 0.392821035  1.05114904
#>    76   11 0.826879867 0.117502109 0.758941593 0.182946456  0.82012889
#>    76   92 0.826879867 0.117502109 0.515737018 0.029036693  0.85134876
#>    14   19 0.986262629 0.262677277 0.861753605 0.258584744  0.92225300
#>    14   76 0.986262629 0.262677277 0.826879867 0.117502109  1.71585740
#>    59    1 0.234462308 0.841599997 0.296339956 0.783732159  0.26729968
#>    12   53 0.100390918 0.955169841 0.163978197 0.805819642  0.16360350
#>    55   52 0.327012023 0.822797390 0.476392917 0.904419636  0.39572766
#>    69   97 0.039924261 0.673309648 0.185735192 0.739555165  0.09468738
#>    35    3 0.262968802 0.704142502 0.215795275 0.606955652  0.33653407
#>    18   50 0.010371495 0.809890943 0.045728987 0.825450735  0.03865709
#>    34   94 0.720784139 0.951154354 0.418434738 1.000000000  0.62611483
#>    54   20 0.323044573 0.230058808 0.281685325 0.342672148  0.34500672
#>    13   15 0.858055326 0.774460188 0.841522451 0.787936158  0.81982210
#>    90    8 0.813417588 0.596557867 0.793567461 0.488653736  0.73671917
#>    16   54 0.222709033 0.213121496 0.323044573 0.230058808  0.27608607
#>   100   98 0.376790209 0.141746597 0.472788951 0.160676808  0.42045622
#>    57   91 0.768855605 0.328764965 0.721106931 0.416219379  0.72221944
#>    40   33 0.641735341 0.204503902 0.610827860 0.316535263  0.59041348
#>    88   34 0.983195595 0.898836683 0.720784139 0.951154354  0.87649979
#>    14   10 0.986262629 0.262677277 1.000000000 0.594264934  1.16168339
#>    11   56 0.758941593 0.182946456 0.762522784 0.255334067  0.75299978
#>    41   62 0.631651753 0.152998736 0.690607266 0.205501831  0.66682781
#>    41   46 0.631651753 0.152998736 0.607083212 0.102881328  0.62697936
#>    58   21 0.206528194 0.082423443 0.171531035 0.020286502  0.23589280
#>    24   92 0.393375525 0.004599453 0.515737018 0.029036693  0.45211523
#>    63   24 0.127727882 0.000000000 0.393375525 0.004599453  0.28112325
#>          my1         mx2         my2 bp1 bp2
#>   0.80413069  0.42254379  0.82546578   0   0
#>   0.81996748  0.31057312  0.80413069   0   0
#>   0.81469618  0.29040272  0.81996748   0   0
#>   0.72241874  0.31057312  0.80413069   0   0
#>   0.75494327  0.22300772  0.70421818   0   0
#>   0.78460385  0.21550355  0.79163318   0   0
#>   0.72776409  0.11481570  0.73030704   0   0
#>   0.77936877  0.09468738  0.74636495   0   0
#>   0.79163318  0.16360350  0.89387166   0   0
#>   0.80421957  0.33947359  0.71573402   0   0
#>   0.71573402  0.33097586  0.72241874   0   0
#>   0.96379233  0.26729968  0.81469618   0   0
#>   0.88583598  0.29040272  0.81996748   0   0
#>   0.87454330  0.41358019  0.92637630   0   0
#>   0.70421818  0.22157603  0.69523967   0   0
#>   0.71532935  0.22157603  0.69523967   0   0
#>   0.65795199  0.14844700  0.71532935   0   0
#>   0.60839245  0.35034260  0.57878622   0   0
#>   0.66257083  0.22157603  0.69523967   0   0
#>   1.17196431 -0.20254850  4.25493098   0   1
#>   0.95064530  0.62611483  1.32533960   0   0
#>   0.92637630  0.54010635  0.98403198   0   0
#>   0.79704130  0.43904800  0.80421957   0   0
#>   0.78909802  0.46304222  0.79704130   0   0
#>   0.67007469  0.64658551  0.62354287   0   0
#>   0.87290638 -0.01165804  0.92600934   0   0
#>   0.96951073  0.33582490  0.88583598   0   0
#>   0.79356823 -0.01165804  0.92600934   0   0
#>   0.74744556  0.11436144  0.87290638   0   0
#>   0.57878622  0.34599126  0.55089016   0   0
#>   0.62370779  0.60309681  0.53384304   0   0
#>   0.74797569 -0.10440945  0.56840274   0   0
#>   0.73414648  0.11408910  0.60363354   0   0
#>   3.07104630 -0.01165804  0.92600934   0   0
#>   0.60363354  0.21371824  0.65795199   0   0
#>   0.53411006  0.10948625  0.59143370   0   0
#>   0.59143370  0.11408910  0.60363354   0   0
#>   0.75697196 -9.92682149  1.39258512   0   1
#>   0.55089016  0.33698818  0.54591200   0   0
#>   0.55023780  0.46893663  0.36792226   0   0
#>   0.56292385  0.10948625  0.59143370   0   0
#>   0.41183366  0.28232563  0.38963688   0   0
#>   0.45789704  0.28577077  0.41183366   0   0
#>   0.54591200  0.28752344  0.53411006   0   0
#>   1.02416033  0.54010635  0.98403198   0   0
#>   0.79600336  0.47577095  0.78909802   0   0
#>   0.76299974  0.64189965  0.79600336   0   0
#>   0.42310598  0.07909967  0.56292385   0   0
#>   0.42272900  0.19989683  0.45789704   0   0
#>   0.30202643  0.23700138  0.29747349   0   0
#>   0.38963688  0.39150153  0.32143873   0   0
#>   0.55810309  0.62273181  0.52787219   0   0
#>   0.57578695  0.83109292  0.66752696   0   0
#>   0.82978088  0.58664456  1.02416033   0   0
#>   0.36633991  0.28232563  0.38963688   0   0
#>   0.40158572  0.13537761  0.42310598   0   0
#>   0.40694566  0.06562356  0.29935146   0   0
#>   0.86882002  0.65616171  0.82978088   0   0
#>   0.74443368  0.79998030  0.83981572   0   0
#>   0.27812502  0.23700138  0.29747349   0   0
#>   0.40100280  0.11648358  0.40158572   0   0
#>   0.68681156  0.81946272  0.72207322   0   0
#>   0.31560991  0.36888304  0.28506695   0   0
#>   0.81906690  0.88089871  0.81905795   0   0
#>   0.72207322  0.65202011  0.76299974   0   0
#>   0.52787219  0.60309681  0.53384304   0   0
#>   0.27224552  0.06562356  0.29935146   0   0
#>   0.29935146  0.11637507  0.40100280   0   0
#>   0.50804173 -1.78460944  0.47790276   0   0
#>   0.53325310  0.72221944  0.36006459   0   0
#>   0.33010803  0.57473861  0.43445238   0   0
#>   0.43445238  0.62273181  0.52787219   0   0
#>   0.57631161  0.84269225  0.57216960   0   0
#>   0.55488948  0.80979375  0.51797309   0   0
#>   0.32143873  0.43327567  0.32914925   0   0
#>   0.26152413  0.11637507  0.40100280   0   0
#>   0.56877775  0.66161150  0.57578695   0   0
#>   0.75553016  0.88089871  0.81905795   0   0
#>   0.57216960  0.91748712  0.57813064   0   0
#>   0.28506695  0.35347715  0.17443233   0   0
#>   0.16380925  0.36888304  0.28506695   0   0
#>   0.32914925  0.41173710  0.18750705   0   0
#>   0.17443233  0.31361450  0.16380925   0   0
#>   0.20257865  0.34500672  0.30202643   0   0
#>   0.81905795  0.92346398  0.87779728   0   0
#>   0.51797309  0.82312932  0.46744651   0   0
#>   0.52778718  0.88536399  0.50719599   0   0
#>   0.41449661  0.79579856  0.41262094   0   0
#>   0.46744651  0.87951824  0.49101563   0   0
#>   0.68668018  0.84990764  0.68681156   0   0
#>   0.18552157  0.53679036  0.12446963   0   0
#>   0.34906440  0.43327567  0.32914925   0   0
#>   0.17318701  0.52487780  0.18552157   0   0
#>   0.41262094  0.86571187  0.41544429   0   0
#>   0.65624657  1.03377109  0.67390483   0   0
#>   0.26224571  0.59041348  0.25062420   0   0
#>   0.23811225  0.05741118  0.27224552   0   0
#>   0.22481051  0.17991403  0.26152413   0   0
#>   0.15884925  0.31173284  0.09112952   0   0
#>   0.12446963  0.59167497  0.14151545   0   0
#>   0.18026688  0.17296586  0.16022853   0   0
#>   1.04793074  0.76703230  0.86882002   0   0
#>   0.78818345  0.91619746  0.75553016   0   0
#>   0.67390483  0.95141568  0.56624856   0   0
#>   0.57813064  0.92130976  0.57261005   0   0
#>   0.27115181  0.66682781  0.17285169   0   0
#>   0.51427766  0.88536399  0.50719599   0   0
#>   0.39787238  0.82757449  0.32663837   0   0
#>   0.43535173  0.87951824  0.49101563   0   0
#>   0.29842909  0.75884377  0.29263988   0   0
#>   0.18278567  0.59041348  0.25062420   0   0
#>   0.28723043  0.70507814  0.26142680   0   0
#>   0.40420925  0.89490014  0.51427766   0   0
#>   0.31672466  0.82757449  0.32663837   0   0
#>   0.76068280  4.52106732  0.94128534   0   1
#>   0.56624856  0.92130976  0.57261005   0   0
#>   0.42148815  0.95141568  0.56624856   0   0
#>   0.08200586  0.04794610  0.18330609   0   0
#>   0.18330609  0.08010473  0.23811225   0   0
#>   0.09867215  0.45211523  0.02904079   0   0
#>   0.14001863  0.36754084  0.07105165   0   0
#>   0.07134067  0.16923097  0.14228776   0   0
#>   0.21952280  0.70507814  0.26142680   0   0
#>   0.17709484  0.73621876  0.18938319   0   0
#>   0.14228776  0.17296586  0.16022853   0   0
#>   0.12420854  0.66272087  0.03110653   0   0
#>   0.14151545  0.61608680  0.18278567   0   0
#>   0.15793739  0.69503178  0.14379716   0   0
#>   0.16244335  0.62697936  0.12420854   0   0
#>   0.21999581  0.84187788  0.31672466   0   0
#>   0.21653156  0.82224684  0.20459145   0   0
#>   0.22800802  0.82224684  0.20459145   0   0
#>   0.31741610  0.86571187  0.41544429   0   0
#>   0.06885115  0.31173284  0.09112952   0   0
#>   0.02496039  0.22175458 -0.14559122   0   0
#>   0.08527881  0.16923097  0.14228776   0   0
#>   0.03110653  0.69503178  0.14379716   0   0
#>  -0.06091542  0.85134876 -0.55995271   0   0
#>   0.10616075  0.66397592 -0.06091542   0   0
#>   0.12215966  0.69503178  0.14379716   0   0
#>   0.13215453  0.73621876  0.18938319   0   0
#>  -1.18583789  0.34934443  0.06885115   0   0
#>   0.07756850  0.12247579  0.08200586   0   0
#>  -1.34903796  0.04794610  0.18330609   0   0
#>   0.40013305  0.89524253  0.43535173   0   0
#>   0.17847966  0.82224684  0.20459145   0   0
#>   0.31402783  0.91014692  0.31741610   0   0
#>  -0.69839739  0.82012889  0.17847966   0   0
#>   0.72241874  0.25340539  0.75494327   0   0
#>   0.78460385  0.25340539  0.75494327   0   0
#>   0.72776409  0.22300772  0.70421818   0   0
#>   0.77936877  0.19483126  0.72776409   0   0
#>   0.79163318  0.19520587  0.77936877   0   0
#>   0.71573402  0.33653407  0.60839245   0   0
#>   0.89387166  0.22283224  0.96379233   0   0
#>   0.96379233  0.22601471  0.96951073   0   0
#>   0.88583598  0.39572766  0.87454330   0   0
#>   0.71532935  0.11481570  0.73030704   0   0
#>   0.74636495  0.09353312  0.74744556   0   0
#>   0.73030704  0.10023854  0.73414648   0   0
#>   0.82546578  0.43904800  0.80421957   0   0
#>   0.66257083  0.21371824  0.65795199   0   0
#>   1.17196431  0.22601471  0.96951073   0   0
#>   0.95064530  0.23201390  1.17196431   0   0
#>   0.92637630  0.40238668  0.95064530   0   0
#>   0.98403198  0.46304222  0.79704130   0   0
#>   0.78909802  0.47609991  0.67007469   0   0
#>   0.79356823 -3.47658075  3.07104630   0   0
#>   0.74744556  0.07963842  0.74797569   0   0
#>   0.57878622  0.43495682  0.62370779   0   0
#>   0.62370779  0.47609991  0.67007469   0   0
#>   3.07104630 -8.45027502  6.15290538   0   1
#>   0.53411006  0.19989683  0.45789704   0   0
#>   0.56840274 -0.62819209  0.50804173   0   0
#>   0.75697196  0.07963842  0.74797569   0   0
#>   0.53384304  0.34660756  0.55023780   0   0
#>   0.55089016  0.34660756  0.55023780   0   0
#>   0.56292385 -0.10440945  0.56840274   0   0
#>   0.41183366  0.33698818  0.54591200   0   0
#>   1.32533960  0.58664456  1.02416033   0   0
#>   0.62354287  0.66161150  0.57578695   0   0
#>   0.76299974  0.64658551  0.62354287   0   0
#>   0.42272900  0.13537761  0.42310598   0   0
#>   0.55810309  0.67654985  0.53325310   0   0
#>   0.82978088  0.64189965  0.79600336   0   0
#>   0.36633991  0.14644256  0.42272900   0   0
#>   0.40158572 -0.03415874  0.40694566   0   0
#>   0.86882002  0.79998030  0.83981572   0   0
#>   0.74443368  0.81946272  0.72207322   0   0
#>   0.27812502  0.17991403  0.26152413   0   0
#>   0.68681156  0.83109292  0.66752696   0   0
#>   0.31560991  0.39150153  0.32143873   0   0
#>   0.81906690  0.79998030  0.83981572   0   0
#>   0.29747349  0.22576919  0.36633991   0   0
#>   0.36792226  0.46273283  0.34906440   0   0
#>   0.27224552 -1.78460944  0.47790276   0   0
#>   0.50804173 -0.03415874  0.40694566   0   0
#>   0.53325310  0.79526344  0.41449661   0   0
#>   0.33010803  0.69171619  0.29842909   0   0
#>   0.43445238  0.46893663  0.36792226   0   0
#>   0.57631161  0.83109292  0.66752696   0   0
#>   0.55488948  0.83634375  0.57631161   0   0
#>   0.56877775  0.65648456  0.55810309   0   0
#>   0.75553016  0.86819510  0.68668018   0   0
#>   0.16380925  0.30546447  0.15884925   0   0
#>   0.17443233  0.41173710  0.18750705   0   0
#>   0.20257865  0.22412155  0.27812502   0   0
#>   0.51797309  0.84200624  0.52778718   0   0
#>   0.52778718  0.84269225  0.57216960   0   0
#>   0.41449661  0.82312932  0.46744651   0   0
#>   0.68668018  0.90213162  0.65624657   0   0
#>   0.18552157  0.53528138  0.26224571   0   0
#>   0.34906440  0.53528138  0.26224571   0   0
#>   0.17318701  0.41173710  0.18750705   0   0
#>   0.50719599  0.87951824  0.49101563   0   0
#>   0.36006459  0.68926457  0.33010803   0   0
#>   0.22481051  0.08010473  0.23811225   0   0
#>   0.47790276 -3.81079136  0.49439797   0   1
#>   0.18026688  0.30546447  0.15884925   0   0
#>   1.04793074  0.92346398  0.87779728   0   0
#>   0.87779728  1.06711732  0.78818345   0   0
#>   0.78818345  1.24773325  0.76068280   0   0
#>   0.16022853  0.14369758  0.22481051   0   0
#>   0.67390483  1.24773325  0.76068280   0   0
#>   0.57813064  0.90213162  0.65624657   0   0
#>   0.27115181  0.68719810  0.28723043   0   0
#>   0.39787238  0.79579856  0.41262094   0   0
#>   0.57261005  0.89490014  0.51427766   0   0
#>   0.43535173  0.86571187  0.41544429   0   0
#>   0.29263988  0.81334908  0.21999581   0   0
#>   0.28723043  0.69171619  0.29842909   0   0
#>   0.32663837  0.75884377  0.29263988   0   0
#>   0.40420925  1.05114904  0.40013305   0   0
#>   0.42148815  1.06350529  0.40420925   0   0
#>   0.09867215  0.52891087  0.10616075   0   0
#>   0.14001863  0.50595757  0.09867215   0   0
#>   0.09112952  0.27249762  0.07134067   0   0
#>   0.07134067  0.23589280  0.02496039   0   0
#>   0.21952280  0.73621876  0.18938319   0   0
#>   0.26142680  0.70449839  0.17709484   0   0
#>   0.17285169  0.61608680  0.18278567   0   0
#>   0.15793739  0.70449839  0.17709484   0   0
#>   0.16244335  0.68702293  0.15793739   0   0
#>   0.21653156  0.81334908  0.21999581   0   0
#>   0.22800802  1.71585740 -0.69839739   0   0
#>   0.31741610  0.84187788  0.31672466   0   0
#>   0.07105165  0.34934443  0.06885115   0   0
#>   0.08527881  0.12247579  0.08200586   0   0
#>  -0.06091542  0.66272087  0.03110653   0   0
#>   0.10616075  0.53679036  0.12446963   0   0
#>   0.02904079  0.36754084  0.07105165   0   0
#>   0.12215966  0.85134876 -0.55995271   0   0
#>   0.13215453  0.77029660  0.12215966   0   0
#>  -1.18583789  0.22175458 -0.14559122   0   0
#>  -0.14559122  0.11840289  0.07756850   0   0
#>   0.07756850 -2.40753602 -1.34903796   0   0
#>  -1.34903796 -5.83309107 -3.36902641   0   1
#>   0.40013305  0.92225300  0.31402783   0   0
#>   0.17847966  0.77550429  0.13215453   0   0
#>  -0.55995271  1.69680744 -3.53352626   0   1
#>   0.31402783  0.92508041  0.22800802   0   0
#>  -0.69839739  4.83933517 -4.12755486   0   1
#>   0.81469618  0.23915740  0.78460385   0   0
#>   0.89387166  0.11436144  0.87290638   0   0
#>   0.87454330  0.42254379  0.82546578   0   0
#>   0.74636495  0.10023854  0.73414648   0   0
#>   0.60839245  0.22491584  0.66257083   0   0
#>   0.79356823  0.05476203  0.75697196   0   0
#>   1.32533960  1.14685145  4.54864443   0   1
#>   0.30202643  0.38199195  0.31560991   0   0
#>   0.74443368  0.88065574  0.81906690   0   0
#>   0.55488948  0.66122333  0.56877775   0   0
#>   0.20257865  0.27985245  0.18026688   0   0
#>   0.17318701  0.42699677  0.14001863   0   0
#>   0.36006459  0.79146656  0.39787238   0   0
#>   0.25062420  0.66482059  0.27115181   0   0
#>   1.04793074  1.60722505  4.71305334   0   1
#>   0.42148815  4.17230984  0.29676067   0   1
#>   0.21952280  0.81346256  0.21653156   0   0
#>   0.17285169  0.67609700  0.16244335   0   0
#>   0.12420854  0.59167497  0.14151545   0   0
#>   0.02496039  0.12879834  0.08527881   0   0
#>   0.02904079  2.02167224 -7.83000320   0   1
#>  -1.18583789  0.34628070 -4.94909475   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.296339956 0.783732159
#>   [2,] 0.484100448 0.518068191
#>   [3,] 0.215795275 0.606955652
#>   [4,] 0.004017651 0.710111042
#>   [5,] 0.487077318 0.564640074
#>   [6,] 0.513027197 0.613193178
#>   [7,] 0.536804302 0.700308919
#>   [8,] 0.793567461 0.488653736
#>   [9,] 0.033408510 0.455071852
#>  [10,] 1.000000000 0.594264934
#>  [11,] 0.758941593 0.182946456
#>  [12,] 0.100390918 0.955169841
#>  [13,] 0.858055326 0.774460188
#>  [14,] 0.986262629 0.262677277
#>  [15,] 0.841522451 0.787936158
#>  [16,] 0.222709033 0.213121496
#>  [17,] 0.860875565 0.547857460
#>  [18,] 0.010371495 0.809890943
#>  [19,] 0.861753605 0.258584744
#>  [20,] 0.281685325 0.342672148
#>  [21,] 0.171531035 0.020286502
#>  [22,] 0.965742888 0.620471102
#>  [23,] 0.716380395 0.104371984
#>  [24,] 0.393375525 0.004599453
#>  [25,] 0.256410586 0.436727306
#>  [26,] 0.881236228 0.173027045
#>  [27,] 0.734058789 0.165865257
#>  [28,] 0.823766229 0.548431046
#>  [29,] 0.179238057 0.695582525
#>  [30,] 0.942582717 0.510865056
#>  [31,] 0.280035323 0.855192000
#>  [32,] 0.171978083 0.687323736
#>  [33,] 0.610827860 0.316535263
#>  [34,] 0.720784139 0.951154354
#>  [35,] 0.262968802 0.704142502
#>  [36,] 0.582485723 0.182081028
#>  [37,] 0.838375337 0.500479961
#>  [38,] 0.324250271 0.410812809
#>  [39,] 0.357432673 0.214384919
#>  [40,] 0.641735341 0.204503902
#>  [41,] 0.631651753 0.152998736
#>  [42,] 0.857347999 0.455088034
#>  [43,] 0.013685040 0.339178137
#>  [44,] 0.468433844 0.197546142
#>  [45,] 0.000000000 0.216604650
#>  [46,] 0.607083212 0.102881328
#>  [47,] 0.762007821 0.691532787
#>  [48,] 0.860571939 0.375255213
#>  [49,] 0.029817269 0.354139084
#>  [50,] 0.045728987 0.825450735
#>  [51,] 0.130932456 0.303655266
#>  [52,] 0.476392917 0.904419636
#>  [53,] 0.163978197 0.805819642
#>  [54,] 0.323044573 0.230058808
#>  [55,] 0.327012023 0.822797390
#>  [56,] 0.762522784 0.255334067
#>  [57,] 0.768855605 0.328764965
#>  [58,] 0.206528194 0.082423443
#>  [59,] 0.234462308 0.841599997
#>  [60,] 0.238489842 0.128945044
#>  [61,] 0.899323911 0.392821035
#>  [62,] 0.690607266 0.205501831
#>  [63,] 0.127727882 0.000000000
#>  [64,] 0.360587265 0.213945642
#>  [65,] 0.718778761 0.205308164
#>  [66,] 0.917209716 0.471998148
#>  [67,] 0.185004849 0.716574391
#>  [68,] 0.782573553 0.775673342
#>  [69,] 0.039924261 0.673309648
#>  [70,] 0.954065857 0.707521436
#>  [71,] 0.180271346 0.326131407
#>  [72,] 0.350568913 0.947756456
#>  [73,] 0.415229328 0.699972898
#>  [74,] 0.793632704 0.278676037
#>  [75,] 0.861397765 0.865215218
#>  [76,] 0.826879867 0.117502109
#>  [77,] 0.438222683 0.988184714
#>  [78,] 0.186767409 0.516800182
#>  [79,] 0.056788456 0.125604826
#>  [80,] 0.298606407 0.009752332
#>  [81,] 0.436101951 0.090614382
#>  [82,] 0.656035908 0.121674086
#>  [83,] 0.248792157 0.422524602
#>  [84,] 0.285330717 0.256195776
#>  [85,] 0.204097553 0.739421867
#>  [86,] 0.100401300 0.157692140
#>  [87,] 0.332724529 0.792817060
#>  [88,] 0.983195595 0.898836683
#>  [89,] 0.856795203 0.599054909
#>  [90,] 0.813417588 0.596557867
#>  [91,] 0.721106931 0.416219379
#>  [92,] 0.515737018 0.029036693
#>  [93,] 0.012199727 0.812712780
#>  [94,] 0.418434738 1.000000000
#>  [95,] 0.529225247 0.882645942
#>  [96,] 0.695496899 0.409466047
#>  [97,] 0.185735192 0.739555165
#>  [98,] 0.472788951 0.160676808
#>  [99,] 0.930833492 0.814898842
#> [100,] 0.376790209 0.141746597
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,87,55), [23,2,4]
#> 2: (31,1,55), [1,14,3]
#> 3: (59,1,31), [2,13,6]
#> 4: (35,87,1), [1,5,11]
#> 5: (1,85,35), [16,4,6]
#> 6: (1,59,85), [9,5,3]
#> 7: (85,97,67), [22,16,8]
#> 8: (85,53,97), [20,7,9]
#> 9: (85,59,53), [12,8,6]
#> 10: (52,87,73), [11,29,23]
#> 11: (73,87,35), [4,21,10]
#> 12: (12,53,59), [9,13,32]
#> 13: (12,59,31), [3,33,12]
#> 14: (72,31,55), [2,15,33]
#> 15: (55,52,72), [27,14,23]
#> 16: (85,67,35), [19,5,7]
#> 17: (32,29,67), [19,22,18]
#> 18: (3,29,32), [17,42,24]
#> 19: (29,35,67), [16,17,24]
#> 20: (69,97,53), [8,35,40]
#> 21: (35,3,73), [36,11,24]
#> 22: (32,67,97), [7,40,17]
#> 23: (52,55,87), [1,10,15]
#> 24: (3,35,29), [19,18,21]
#> 25: (72,94,12), [0,33,26]
#> 26: (72,77,94), [54,25,27]
#> 27: (72,52,77), [28,26,15]
#> 28: (95,77,52), [27,29,55]
#> 29: (95,52,73), [10,30,28]
#> 30: (7,95,73), [29,31,56]
#> 31: (73,6,7), [57,30,37]
#> 32: (53,12,50), [39,35,12]
#> 33: (12,31,72), [14,25,13]
#> 34: (18,50,93), [39,41,46]
#> 35: (69,53,50), [32,38,20]
#> 36: (73,3,5), [48,37,21]
#> 37: (73,5,6), [47,31,36]
#> 38: (50,4,69), [45,35,46]
#> 39: (93,50,12), [32,41,34]
#> 40: (97,69,32), [42,22,20]
#> 41: (18,93,12), [39,0,34]
#> 42: (69,3,32), [18,40,44]
#> 43: (25,3,78), [44,52,53]
#> 44: (78,3,69), [42,50,43]
#> 45: (9,69,4), [38,82,50]
#> 46: (50,18,4), [0,38,34]
#> 47: (2,6,5), [37,49,77]
#> 48: (5,3,38), [53,49,36]
#> 49: (5,38,2), [79,47,48]
#> 50: (9,78,69), [44,45,59]
#> 51: (25,83,38), [62,53,52]
#> 52: (78,83,25), [51,43,60]
#> 53: (38,3,25), [43,51,48]
#> 54: (34,94,77), [26,55,0]
#> 55: (34,77,95), [28,65,54]
#> 56: (68,95,7), [30,58,65]
#> 57: (47,7,6), [31,64,58]
#> 58: (47,68,7), [56,57,76]
#> 59: (71,78,9), [50,67,60]
#> 60: (71,83,78), [52,59,66]
#> 61: (54,20,84), [78,99,74]
#> 62: (83,20,38), [89,51,66]
#> 63: (8,6,96), [77,84,91]
#> 64: (6,90,47), [93,57,91]
#> 65: (68,34,95), [55,56,69]
#> 66: (71,20,83), [62,60,78]
#> 67: (49,71,9), [59,68,72]
#> 68: (9,43,49), [81,67,82]
#> 69: (75,34,68), [65,83,123]
#> 70: (13,15,68), [83,76,75]
#> 71: (16,84,71), [78,90,99]
#> 72: (51,71,49), [67,81,90]
#> 73: (89,13,47), [76,93,105]
#> 74: (20,54,64), [95,89,61]
#> 75: (15,13,75), [100,83,70]
#> 76: (13,68,47), [58,73,70]
#> 77: (96,6,2), [47,86,63]
#> 78: (71,84,20), [61,66,71]
#> 79: (33,2,38), [49,107,86]
#> 80: (45,51,43), [81,118,116]
#> 81: (43,51,49), [72,68,80]
#> 82: (9,4,43), [118,68,45]
#> 83: (75,68,15), [70,75,69]
#> 84: (8,96,91), [113,103,63]
#> 85: (57,96,33), [86,135,113]
#> 86: (33,96,2), [77,79,85]
#> 87: (90,28,89), [94,93,88]
#> 88: (90,8,28), [101,87,91]
#> 89: (20,64,38), [97,62,74]
#> 90: (16,71,51), [72,117,71]
#> 91: (8,90,6), [64,63,88]
#> 92: (70,99,13), [100,105,125]
#> 93: (89,47,90), [64,87,73]
#> 94: (28,17,89), [128,87,102]
#> 95: (54,39,64), [98,74,96]
#> 96: (100,39,54), [95,119,98]
#> 97: (38,64,44), [115,107,89]
#> 98: (64,39,100), [96,115,95]
#> 99: (16,54,84), [61,71,122]
#> 100: (13,99,75), [124,75,92]
#> 101: (28,8,37), [104,102,88]
#> 102: (28,37,17), [110,94,101]
#> 103: (8,91,42), [111,104,84]
#> 104: (8,42,37), [109,101,103]
#> 105: (70,13,89), [73,112,92]
#> 106: (44,98,36), [121,114,108]
#> 107: (33,38,44), [97,114,79]
#> 108: (100,98,44), [106,115,148]
#> 109: (66,37,42), [104,134,110]
#> 110: (66,17,37), [102,109,130]
#> 111: (91,48,42), [131,103,132]
#> 112: (89,22,70), [127,105,128]
#> 113: (57,91,96), [84,85,132]
#> 114: (44,36,33), [120,107,106]
#> 115: (100,44,64), [97,98,108]
#> 116: (86,51,45), [80,146,117]
#> 117: (86,16,51), [90,116,126]
#> 118: (45,43,4), [82,0,80]
#> 119: (54,60,100), [149,96,122]
#> 120: (40,33,36), [114,137,129]
#> 121: (98,46,36), [157,106,171]
#> 122: (54,16,60), [126,119,99]
#> 123: (88,34,75), [69,124,0]
#> 124: (88,75,99), [100,125,123]
#> 125: (88,99,70), [92,142,124]
#> 126: (86,60,16), [122,117,155]
#> 127: (70,22,10), [143,142,112]
#> 128: (89,17,22), [133,112,94]
#> 129: (33,40,62), [153,138,120]
#> 130: (30,17,66), [110,140,133]
#> 131: (61,42,48), [111,164,134]
#> 132: (91,57,48), [139,111,113]
#> 133: (30,22,17), [128,130,143]
#> 134: (61,66,42), [109,131,180]
#> 135: (33,56,57), [136,85,138]
#> 136: (74,57,56), [135,161,139]
#> 137: (41,40,36), [120,157,153]
#> 138: (33,62,56), [152,135,129]
#> 139: (74,48,57), [132,136,141]
#> 140: (14,30,66), [130,180,144]
#> 141: (19,48,74), [139,161,164]
#> 142: (70,10,88), [0,125,127]
#> 143: (10,22,30), [133,144,127]
#> 144: (14,10,30), [143,140,0]
#> 145: (21,86,79), [146,178,168]
#> 146: (79,86,45), [116,179,145]
#> 147: (98,81,92), [172,171,148]
#> 148: (98,100,81), [166,147,108]
#> 149: (80,100,60), [119,150,165]
#> 150: (80,60,58), [155,167,149]
#> 151: (11,56,65), [152,158,162]
#> 152: (65,56,62), [138,154,151]
#> 153: (41,62,40), [129,137,160]
#> 154: (62,27,65), [158,152,159]
#> 155: (58,60,86), [126,168,150]
#> 156: (41,46,82), [169,160,157]
#> 157: (46,41,36), [137,121,156]
#> 158: (11,65,27), [154,174,151]
#> 159: (62,82,27), [175,154,160]
#> 160: (62,41,82), [156,159,153]
#> 161: (56,19,74), [141,136,162]
#> 162: (56,11,19), [181,161,151]
#> 163: (14,19,26), [181,185,184]
#> 164: (19,61,48), [131,141,184]
#> 165: (24,100,80), [149,176,166]
#> 166: (24,81,100), [148,165,172]
#> 167: (58,21,80), [177,150,168]
#> 168: (21,58,86), [155,145,167]
#> 169: (46,23,82), [175,156,170]
#> 170: (46,92,23), [183,169,171]
#> 171: (98,92,46), [170,121,147]
#> 172: (24,92,81), [147,166,0]
#> 173: (76,27,23), [175,183,174]
#> 174: (76,11,27), [158,173,182]
#> 175: (82,23,27), [173,159,169]
#> 176: (63,24,80), [165,177,0]
#> 177: (63,80,21), [167,178,176]
#> 178: (63,21,79), [145,179,177]
#> 179: (63,79,45), [146,0,178]
#> 180: (14,66,61), [134,184,140]
#> 181: (11,26,19), [163,162,182]
#> 182: (76,26,11), [181,174,185]
#> 183: (76,23,92), [170,0,173]
#> 184: (14,61,19), [164,163,180]
#> 185: (14,26,76), [182,0,163]
#> boundary nodes:  94 12 18 4 45 63 24 92 76 14 10 88 34 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.03051971 0.04966757 0.07329632 0.03750163 0.06863277 0.10659451 0.02299238 
#>                                                                              
#> 0.06974488 0.07904583 0.12420580 0.11282304 0.04755673 0.05706331 0.13307802 
#>                                                                              
#> 0.07894891 0.02176957 0.01099613 0.22005942 0.08416716 0.32118782 0.02304699 
#>                                                                              
#> 0.09205192 0.21339975 0.21532420 0.09030229 0.14076567 0.11637511 0.03586735 
#>                                                                              
#> 0.11986766 0.27456248 0.05505270 0.05141622 0.13279536 0.16754611 0.09153674 
#>                                                                              
#> 0.09471339 0.18797198 0.09998200 0.22413054 0.19249882 0.21461156 0.07636159 
#>                                                                              
#> 0.01611699 0.17500655 0.13935530 0.18249447 0.25706482 0.16531587 0.11284925 
#>                                                                              
#> 0.08655317 0.08034254 0.27349577 0.10799629 0.20344093 0.08636188 0.19543370 
#>                                                                              
#> 0.02200179 0.18604171 0.06021087 0.12620802 0.15303875 0.12689396 0.04085447 
#>                                                                              
#> 0.09081656 0.08661745 0.09942595 0.12251053 0.11301724 0.37105886 0.02648549 
#>                                                                              
#> 0.12571785 0.23766128 0.06044575 0.06697233 0.20019255 0.05421718 0.30085066 
#>                                                                              
#> 0.08325831 0.05135979 0.00318503 0.03779169 0.10908633 0.07517340 0.04588547 
#>                                                                              
#> 0.08574997 0.04634226 0.05244891 0.14167710 0.04919745 0.17540981 0.11176548 
#>                                                                              
#> 0.25743266 0.03712566 0.07989789 0.08783003 0.13740894 0.15722951 0.12891588 
#>                                                                              
#> 0.13889158 0.08293147 0.08564292 0.16479602 0.10986196 0.04313135 0.12755350 
#>                                                                              
#> 0.04888211 0.09448900 0.10282621 0.06220432 0.07370347 0.06335063 0.08749339 
#>                                                                              
#> 0.04641579 0.11750914 0.30503498 0.11202624 0.10125356 0.05414495 0.11640921 
#>                                                                              
#> 0.10066556 0.07830955 0.05644295 0.06645397 0.04229920 0.14104907 0.05243600 
#>                                                                              
#> 0.05712336 0.08966844 0.03969661 0.07102195 0.12763800 0.08774789 0.04254731 
#>                                                                              
#> 0.15341184 0.12751125 0.13010878 0.06277595 0.21432025 0.11746130 0.06398397 
#>                                                                              
#> 0.03018147 0.09490910 0.15574919 0.10726554 0.08117215 0.12269626 0.13939480 
#>                                                                              
#> 0.07770220 0.08630265 0.10233310 0.02977480 0.01836284 0.07757723 0.15231761 
#>                                                                              
#> 0.17570785 0.20559105 0.12716012 0.03202027 0.18151660 0.05401278 0.18192235 
#>                                                                              
#> 0.09587050 0.25028781 0.08564554 0.09652786 0.05714323 0.12157544 0.00336232 
#>                                                                              
#> 0.15225178 0.15322241 0.13074840 0.17090774 0.10612181 0.25672711 0.12265024 
#>                                                                              
#> 0.04666693 0.22399883 0.21833504 0.07262082 0.28497759 0.26101426 0.22537446 
#>                                                                              
#> 0.19077940 0.12604969 0.27500637 0.11826559 0.10099664 0.11929377 0.07549152 
#>                                                                              
#> 0.12071536 0.13242661 0.15098354 0.07979399 0.10275403 0.30168687 0.12333507 
#>                                                                              
#> 0.11756006 0.10245616 0.15850026 0.23806587 0.04344943 0.04922688 0.30694122 
#>                                                                              
#> 0.11704195 0.10338104 0.07399485 0.07600572 0.05012716 0.03711377 0.07207367 
#>                                                                              
#> 0.14569340 0.11509562 0.18556526 0.10729466 0.08382167 0.10906037 0.14912205 
#>                                                                              
#> 0.49352275 0.13180856 0.12635313 0.09893104 0.19352019 0.13428177 0.12221694 
#>                                                                              
#> 0.11103266 0.13672299 0.14535664 0.08969112 0.07509432 0.03889312 0.16357548 
#>                                                                              
#> 0.05588207 0.22041672 0.25200220 0.13846897 0.07908653 0.13349496 0.11730084 
#>                                                                              
#> 0.04596846 0.02817216 0.05248296 0.05881405 0.09067673 0.09928405 0.13808588 
#>                                                                              
#> 0.11667645 0.13814634 0.15472475 0.10930735 0.14620282 0.09604226 0.11127683 
#>                                                                              
#> 0.10466493 0.17115659 0.04827275 0.14425316 0.25145971 0.15651114 0.09433224 
#>                                                                              
#> 0.32347489 0.12457627 0.21558918 0.08472030 0.16232321 0.17022586 0.16015398 
#>                                                                              
#> 0.10803067 0.03862977 0.30626958 0.11996813 0.02132927 0.10971476 0.10175506 
#>                                                                              
#> 0.09784739 0.09964041 0.11621660 0.26757599 0.33187210 0.07247614 0.07894509 
#>                                             
#> 0.05581548 0.07131480 0.12477786 0.26568746 
#> 
#> $alpha
#> [1] 0.2693745
#> 
#> attr(,"class")
#> [1] "scree"