Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1          y1         x2          y2         mx1
#>    16   70 0.49283865 0.679207890 0.57934396 0.710410992  0.53062674
#>     1   70 0.51831685 0.757097527 0.57934396 0.710410992  0.56752767
#>    86   70 0.58410591 0.804542001 0.57934396 0.710410992  0.57455050
#>    16   93 0.49283865 0.679207890 0.40146469 0.633799017  0.47808225
#>   100    3 0.40968833 0.856293847 0.36719013 0.826663324  0.43017181
#>     3    1 0.36719013 0.826663324 0.51831685 0.757097527  0.43273863
#>     1   86 0.51831685 0.757097527 0.58410591 0.804542001  0.52326355
#>     6   70 0.61712288 0.779287687 0.57934396 0.710410992  0.65120505
#>    16   35 0.49283865 0.679207890 0.52314194 0.520758748  0.56627501
#>    35   15 0.52314194 0.520758748 0.68540754 0.652365769  0.59006712
#>     3   16 0.36719013 0.826663324 0.49283865 0.679207890  0.42147853
#>    15   31 0.68540754 0.652365769 0.71423987 0.600564108  0.76949552
#>    98    8 0.32417201 0.816224074 0.28682332 0.685418383  0.36708655
#>    79  100 0.42261136 0.944883583 0.40968833 0.856293847  0.44419854
#>    73   79 0.49674651 0.950103171 0.42261136 0.944883583  0.46265296
#>    51   86 0.46802100 0.849171133 0.58410591 0.804542001  0.54697663
#>    86   47 0.58410591 0.804542001 0.69041807 0.896394933  0.62404498
#>    99   78 0.48210467 0.471884766 0.55653073 0.404554679  0.54198273
#>    35   93 0.52314194 0.520758748 0.40146469 0.633799017  0.43772046
#>     8   96 0.28682332 0.685418383 0.33283659 0.552836555  0.32889601
#>     8   93 0.28682332 0.685418383 0.40146469 0.633799017  0.37521380
#>     6   47 0.61712288 0.779287687 0.69041807 0.896394933  0.64157931
#>     6   15 0.61712288 0.779287687 0.68540754 0.652365769  0.74474437
#>    86   11 0.58410591 0.804542001 0.67956508 0.913316393  0.59934079
#>    96   99 0.33283659 0.552836555 0.48210467 0.471884766  0.42485075
#>    18   32 0.56539786 0.966417381 0.44011597 0.971695847  0.50663474
#>    73   86 0.49674651 0.950103171 0.58410591 0.804542001  0.54907291
#>    18   86 0.56539786 0.966417381 0.58410591 0.804542001  0.59612215
#>    98   75 0.32417201 0.816224074 0.36347308 0.955071926  0.32435578
#>    42   11 0.61014115 0.971043767 0.67956508 0.913316393  0.62804932
#>    11   47 0.67956508 0.913316393 0.69041807 0.896394933  0.73334404
#>    79   32 0.42261136 0.944883583 0.44011597 0.971695847  0.41523041
#>    38   32 0.39394805 0.982036406 0.44011597 0.971695847  0.52302435
#>     8   27 0.28682332 0.685418383 0.26251826 0.769107106  0.24695311
#>    75   79 0.36347308 0.955071926 0.42261136 0.944883583  0.39338527
#>    42   38 0.61014115 0.971043767 0.39394805 0.982036406  0.53003740
#>    78   31 0.55653073 0.404554679 0.71423987 0.600564108  0.64678596
#>    71   78 0.38109901 0.409905081 0.55653073 0.404554679  0.46803100
#>    88   36 0.90637839 0.712887409 0.83455066 0.857745918  0.78004664
#>    25   43 0.81788525 0.514251722 0.84626390 0.628593337  0.78704122
#>    96   71 0.33283659 0.552836555 0.38109901 0.409905081  0.39821787
#>    72   71 0.24058390 0.408248479 0.38109901 0.409905081  0.31017542
#>    77   33 0.85865842 0.887969978 0.67555588 0.943695073  0.77741938
#>    84   23 0.23844051 0.667645097 0.20478466 0.617837041  0.27845336
#>    78   41 0.55653073 0.404554679 0.73750297 0.376850088  0.66098726
#>    50   64 0.06933649 0.863553743 0.25808914 0.926141993  0.17410726
#>    27   98 0.26251826 0.769107106 0.32417201 0.816224074  0.25153349
#>    19   64 0.20568461 0.963399359 0.25808914 0.926141993  0.28797931
#>    43   52 0.84626390 0.628593337 0.88178782 0.615568756  0.88071704
#>    77   14 0.85865842 0.887969978 0.91294921 0.938452123  0.84246157
#>    36   77 0.83455066 0.857745918 0.85865842 0.887969978  0.77637569
#>    43   88 0.84626390 0.628593337 0.90637839 0.712887409  0.78041576
#>     4   54 0.16283406 0.619203685 0.14096924 0.608837088  0.18162073
#>    84   27 0.23844051 0.667645097 0.26251826 0.769107106  0.15273777
#>    60   25 0.82805874 0.396482289 0.81788525 0.514251722  0.76889941
#>    41   25 0.73750297 0.376850088 0.81788525 0.514251722  0.70723718
#>    46   41 0.75636274 0.352915334 0.73750297 0.376850088  0.71240330
#>    23   72 0.20478466 0.617837041 0.24058390 0.408248479  0.23300857
#>    10   54 0.09913948 0.641885370 0.14096924 0.608837088  0.13659724
#>    50   27 0.06933649 0.863553743 0.26251826 0.769107106  0.18012143
#>    38   19 0.39394805 0.982036406 0.20568461 0.963399359  0.29034157
#>     7   52 0.92780640 0.557419866 0.88178782 0.615568756  0.86524412
#>    89   88 0.95869074 0.776378352 0.90637839 0.712887409  0.96152308
#>    88   52 0.90637839 0.712887409 0.88178782 0.615568756  0.93810180
#>    46   60 0.75636274 0.352915334 0.82805874 0.396482289  0.78158403
#>    62    7 0.93826972 0.502012923 0.92780640 0.557419866  0.87923893
#>    26   25 0.87576094 0.450803821 0.81788525 0.514251722  0.81642755
#>    89   14 0.95869074 0.776378352 0.91294921 0.938452123  0.97845446
#>    89   92 0.95869074 0.776378352 0.91888130 0.795261778  0.92298235
#>    71   30 0.38109901 0.409905081 0.30828918 0.338024685  0.41210280
#>    10   91 0.09913948 0.641885370 0.02944372 0.549120584  0.08448046
#>     4   84 0.16283406 0.619203685 0.23844051 0.667645097  0.15881689
#>     9   50 0.17879399 0.969079531 0.06933649 0.863553743  0.17133669
#>    71   21 0.38109901 0.409905081 0.53860805 0.264155914  0.46690636
#>    61   42 0.90419539 1.000000000 0.61014115 0.971043767  0.73761887
#>    21   68 0.53860805 0.264155914 0.69073715 0.296500699  0.60473732
#>    65   67 0.53672626 0.255558464 0.63585357 0.204063228  0.59534087
#>     2   65 0.31344089 0.259979208 0.53672626 0.255558464  0.42561052
#>    62   34 0.93826972 0.502012923 1.00000000 0.680471016  1.07144148
#>    26   62 0.87576094 0.450803821 0.93826972 0.502012923  0.87840300
#>    45   62 0.91163184 0.431537071 0.93826972 0.502012923  0.91040246
#>    34   61 1.00000000 0.680471016 0.90419539 1.000000000  1.89599796
#>    60   45 0.82805874 0.396482289 0.91163184 0.431537071  0.87390695
#>    67   81 0.63585357 0.204063228 0.83458993 0.131957994  0.74703271
#>    61   14 0.90419539 1.000000000 0.91294921 0.938452123  1.76215797
#>    72   63 0.24058390 0.408248479 0.11426554 0.481253969  0.17227374
#>    23   54 0.20478466 0.617837041 0.14096924 0.608837088  0.18417775
#>    50   55 0.06933649 0.863553743 0.04006796 0.666318977  0.11193485
#>    10   27 0.09913948 0.641885370 0.26251826 0.769107106  0.13838734
#>    10   55 0.09913948 0.641885370 0.04006796 0.666318977  0.04985262
#>    72   17 0.24058390 0.408248479 0.23185851 0.296358543  0.25158578
#>    60   40 0.82805874 0.396482289 0.95517798 0.336311050  0.89066309
#>    17    2 0.23185851 0.296358543 0.31344089 0.259979208  0.28106237
#>    63   91 0.11426554 0.481253969 0.02944372 0.549120584  0.02003844
#>     2   87 0.31344089 0.259979208 0.30381473 0.251470547  0.33650320
#>    40   80 0.95517798 0.336311050 0.95917671 0.342260156  0.90802997
#>    67   68 0.63585357 0.204063228 0.69073715 0.296500699  0.61495905
#>    65   29 0.53672626 0.255558464 0.49175409 0.156311258  0.52297983
#>    13    2 0.35700441 0.177564949 0.31344089 0.259979208  0.39306709
#>    13   29 0.35700441 0.177564949 0.49175409 0.156311258  0.43611389
#>    24   87 0.34763999 0.183093163 0.30381473 0.251470547  0.27620991
#>    87   17 0.30381473 0.251470547 0.23185851 0.296358543  0.23191551
#>    34   49 1.00000000 0.680471016 0.96231590 0.455206657  2.19749666
#>    49   45 0.96231590 0.455206657 0.91163184 0.431537071  0.95778167
#>    81   46 0.83458993 0.131957994 0.75636274 0.352915334  0.79455455
#>    58   12 0.55003801 0.153724802 0.46837686 0.089996649  0.56416662
#>    91   55 0.02944372 0.549120584 0.04006796 0.666318977 -0.70677203
#>    57   74 0.15367777 0.364552732 0.07281106 0.366496206  0.11084934
#>    57   63 0.15367777 0.364552732 0.11426554 0.481253969  0.11446509
#>     5   12 0.41586367 0.125379462 0.46837686 0.089996649  0.45782392
#>     5   13 0.41586367 0.125379462 0.35700441 0.177564949  0.36904327
#>    95    5 0.45846331 0.080601040 0.41586367 0.125379462  0.42606531
#>    48   44 0.45782894 0.001918630 0.66739529 0.053972351  0.56105171
#>    67   58 0.63585357 0.204063228 0.55003801 0.153724802  0.62543147
#>    17   57 0.23185851 0.296358543 0.15367777 0.364552732  0.13122043
#>    59   39 0.22962010 0.099653739 0.39163810 0.090341300  0.31143291
#>    24   82 0.34763999 0.183093163 0.22403730 0.136694935  0.29750104
#>    81   28 0.83458993 0.131957994 0.88802002 0.112019023  0.90166271
#>    40   46 0.95517798 0.336311050 0.75636274 0.352915334  0.84881405
#>    44   37 0.66739529 0.053972351 0.66781593 0.052517428  0.75348629
#>    48   95 0.45782894 0.001918630 0.45846331 0.080601040  0.41865300
#>    24   59 0.34763999 0.183093163 0.22962010 0.099653739  0.31080715
#>    82   83 0.22403730 0.136694935 0.09361727 0.159676782  0.15439663
#>    90   80 0.96440074 0.106119471 0.95917671 0.342260156  1.12310188
#>    97   28 0.83576524 0.043106610 0.88802002 0.112019023  0.84851248
#>    56   83 0.01154638 0.101679416 0.09361727 0.159676782 -0.03274405
#>    81   44 0.83458993 0.131957994 0.66739529 0.053972351  0.75402739
#>    85   97 0.81219968 0.000000000 0.83576524 0.043106610  0.75250864
#>    34   80 1.00000000 0.680471016 0.95917671 0.342260156  2.98749684
#>    28   90 0.88802002 0.112019023 0.96440074 0.106119471  0.93479505
#>    85   94 0.81219968 0.000000000 0.89352319 0.032496758  0.85822113
#>    48   85 0.45782894 0.001918630 0.81219968 0.000000000  0.63356691
#>    48   39 0.45782894 0.001918630 0.39163810 0.090341300  0.34787807
#>    59   69 0.22962010 0.099653739 0.12597158 0.058017427  0.16771146
#>    59   82 0.22962010 0.099653739 0.22403730 0.136694935  0.16569455
#>    76   59 0.23787853 0.001510489 0.22962010 0.099653739  0.20093839
#>    69   22 0.12597158 0.058017427 0.02946219 0.061416008  0.07708611
#>    56   22 0.01154638 0.101679416 0.02946219 0.061416008  0.07132169
#>    91   56 0.02944372 0.549120584 0.01154638 0.101679416 -0.43049307
#>    91   53 0.02944372 0.549120584 0.00000000 0.916430928 -3.37364320
#>    20   76 0.02576702 0.037445920 0.23787853 0.001510489  0.10144937
#>    66   69 0.14169986 0.047597221 0.12597158 0.058017427  0.08878675
#>    51    3 0.46802100 0.849171133 0.36719013 0.826663324  0.43017181
#>    51    1 0.46802100 0.849171133 0.51831685 0.757097527  0.43273863
#>    15   70 0.68540754 0.652365769 0.57934396 0.710410992  0.65120505
#>    70   35 0.57934396 0.710410992 0.52314194 0.520758748  0.56627501
#>     1   16 0.51831685 0.757097527 0.49283865 0.679207890  0.42147853
#>    43   31 0.84626390 0.628593337 0.71423987 0.600564108  0.76949552
#>    31   35 0.71423987 0.600564108 0.52314194 0.520758748  0.61169506
#>     3    8 0.36719013 0.826663324 0.28682332 0.685418383  0.36708655
#>    75  100 0.36347308 0.955071926 0.40968833 0.856293847  0.35422837
#>    51  100 0.46802100 0.849171133 0.40968833 0.856293847  0.44419854
#>    51   79 0.46802100 0.849171133 0.42261136 0.944883583  0.46265296
#>    35   78 0.52314194 0.520758748 0.55653073 0.404554679  0.54198273
#>    99   93 0.48210467 0.471884766 0.40146469 0.633799017  0.43772046
#>    93   96 0.40146469 0.633799017 0.33283659 0.552836555  0.32889601
#>     3   93 0.36719013 0.826663324 0.40146469 0.633799017  0.37521380
#>    36   15 0.83455066 0.857745918 0.68540754 0.652365769  0.74474437
#>    73   32 0.49674651 0.950103171 0.44011597 0.971695847  0.50663474
#>    42   86 0.61014115 0.971043767 0.58410591 0.804542001  0.59612215
#>    64   75 0.25808914 0.926141993 0.36347308 0.955071926  0.32435578
#>    33   11 0.67555588 0.943695073 0.67956508 0.913316393  0.62804932
#>    33   47 0.67555588 0.943695073 0.69041807 0.896394933  0.73334404
#>    84    8 0.23844051 0.667645097 0.28682332 0.685418383  0.28667166
#>     3   98 0.36719013 0.826663324 0.32417201 0.816224074  0.32908849
#>    38   79 0.39394805 0.982036406 0.42261136 0.944883583  0.39338527
#>    18   38 0.56539786 0.966417381 0.39394805 0.982036406  0.53003740
#>    36    6 0.83455066 0.857745918 0.61712288 0.779287687  0.73785351
#>    99   71 0.48210467 0.471884766 0.38109901 0.409905081  0.39821787
#>    96   23 0.33283659 0.552836555 0.20478466 0.617837041  0.27845336
#>    31   41 0.71423987 0.600564108 0.73750297 0.376850088  0.66098726
#>    64   98 0.25808914 0.926141993 0.32417201 0.816224074  0.25153349
#>    52   25 0.88178782 0.615568756 0.81788525 0.514251722  0.84433648
#>    47   77 0.69041807 0.896394933 0.85865842 0.887969978  0.77637569
#>    92   88 0.91888130 0.795261778 0.90637839 0.712887409  0.83210591
#>    15   88 0.68540754 0.652365769 0.90637839 0.712887409  0.78041576
#>    31   25 0.71423987 0.600564108 0.81788525 0.514251722  0.70723718
#>    68   41 0.69073715 0.296500699 0.73750297 0.376850088  0.71240330
#>    92   36 0.91888130 0.795261778 0.83455066 0.857745918  0.88711811
#>    96   72 0.33283659 0.552836555 0.24058390 0.408248479  0.23300857
#>    64   27 0.25808914 0.926141993 0.26251826 0.769107106  0.18012143
#>    75   19 0.36347308 0.955071926 0.20568461 0.963399359  0.29034157
#>    30   72 0.30828918 0.338024685 0.24058390 0.408248479  0.31085126
#>    63   23 0.11426554 0.481253969 0.20478466 0.617837041  0.21625946
#>    68   78 0.69073715 0.296500699 0.55653073 0.404554679  0.64491250
#>    34   88 1.00000000 0.680471016 0.90637839 0.712887409  0.96152308
#>    34   52 1.00000000 0.680471016 0.88178782 0.615568756  0.93810180
#>    41   60 0.73750297 0.376850088 0.82805874 0.396482289  0.78158403
#>    25    7 0.81788525 0.514251722 0.92780640 0.557419866  0.87923893
#>    92   14 0.91888130 0.795261778 0.91294921 0.938452123  0.97845446
#>     2   30 0.31344089 0.259979208 0.30828918 0.338024685  0.41210280
#>    54   91 0.14096924 0.608837088 0.02944372 0.549120584  0.08448046
#>    10   84 0.09913948 0.641885370 0.23844051 0.667645097  0.15881689
#>    19   50 0.20568461 0.963399359 0.06933649 0.863553743  0.17133669
#>     9   38 0.17879399 0.969079531 0.39394805 0.982036406  0.26592480
#>    78   21 0.55653073 0.404554679 0.53860805 0.264155914  0.46690636
#>    33   42 0.67555588 0.943695073 0.61014115 0.971043767  0.73761887
#>    92   77 0.91888130 0.795261778 0.85865842 0.887969978  0.92842418
#>    21   67 0.53860805 0.264155914 0.63585357 0.204063228  0.59534087
#>    21   65 0.53860805 0.264155914 0.53672626 0.255558464  0.42561052
#>     7   34 0.92780640 0.557419866 1.00000000 0.680471016  1.07144148
#>    25   62 0.81788525 0.514251722 0.93826972 0.502012923  0.87840300
#>    89   61 0.95869074 0.776378352 0.90419539 1.000000000  1.89599796
#>    26   45 0.87576094 0.450803821 0.91163184 0.431537071  0.87390695
#>    68   81 0.69073715 0.296500699 0.83458993 0.131957994  0.74703271
#>    61   77 0.90419539 1.000000000 0.85865842 0.887969978  0.78731687
#>    63   54 0.11426554 0.481253969 0.14096924 0.608837088  0.18417775
#>     2   71 0.31344089 0.259979208 0.38109901 0.409905081  0.42532609
#>    30   17 0.30828918 0.338024685 0.23185851 0.296358543  0.25158578
#>    49   62 0.96231590 0.455206657 0.93826972 0.502012923  1.08889589
#>    45   40 0.91163184 0.431537071 0.95517798 0.336311050  0.89066309
#>    57   72 0.15367777 0.364552732 0.24058390 0.408248479  0.21337899
#>    74   91 0.07281106 0.366496206 0.02944372 0.549120584  0.02003844
#>    45   80 0.91163184 0.431537071 0.95917671 0.342260156  0.90802997
#>    58   29 0.55003801 0.153724802 0.49175409 0.156311258  0.52297983
#>    24    2 0.34763999 0.183093163 0.31344089 0.259979208  0.39306709
#>    58   65 0.55003801 0.153724802 0.53672626 0.255558464  0.57538832
#>    82   87 0.22403730 0.136694935 0.30381473 0.251470547  0.27620991
#>    82   17 0.22403730 0.136694935 0.23185851 0.296358543  0.23191551
#>    80   49 0.95917671 0.342260156 0.96231590 0.455206657  2.19749666
#>    68   46 0.69073715 0.296500699 0.75636274 0.352915334  0.79455455
#>    44   12 0.66739529 0.053972351 0.46837686 0.089996649  0.56416662
#>    53   55 0.00000000 0.916430928 0.04006796 0.666318977 -0.70677203
#>    53   50 0.00000000 0.916430928 0.06933649 0.863553743 -0.04896926
#>    53    9 0.00000000 0.916430928 0.17879399 0.969079531  0.08536036
#>    83   74 0.09361727 0.159676782 0.07281106 0.366496206  0.11084934
#>    74   63 0.07281106 0.366496206 0.11426554 0.481253969  0.11446509
#>    29   12 0.49175409 0.156311258 0.46837686 0.089996649  0.45782392
#>    39   13 0.39163810 0.090341300 0.35700441 0.177564949  0.36904327
#>    39    5 0.39163810 0.090341300 0.41586367 0.125379462  0.42606531
#>    44   58 0.66739529 0.053972351 0.55003801 0.153724802  0.62543147
#>    83   57 0.09361727 0.159676782 0.15367777 0.364552732  0.13122043
#>    40   28 0.95517798 0.336311050 0.88802002 0.112019023  0.90166271
#>    44   67 0.66739529 0.053972351 0.63585357 0.204063228  0.72680683
#>    97   37 0.83576524 0.043106610 0.66781593 0.052517428  0.75348629
#>    39   95 0.39163810 0.090341300 0.45846331 0.080601040  0.41865300
#>    13   59 0.35700441 0.177564949 0.22962010 0.099653739  0.31080715
#>    69   83 0.12597158 0.058017427 0.09361727 0.159676782  0.15439663
#>    12   48 0.46837686 0.089996649 0.45782894 0.001918630  0.50552025
#>    97   44 0.83576524 0.043106610 0.66739529 0.053972351  0.75402739
#>    40   90 0.95517798 0.336311050 0.96440074 0.106119471  0.93479505
#>    94   28 0.89352319 0.032496758 0.88802002 0.112019023  0.92354228
#>    97   94 0.83576524 0.043106610 0.89352319 0.032496758  0.85822113
#>    37   85 0.66781593 0.052517428 0.81219968 0.000000000  0.63356691
#>    76   48 0.23787853 0.001510489 0.45782894 0.001918630  0.42115206
#>    76   39 0.23787853 0.001510489 0.39163810 0.090341300  0.34787807
#>    69   82 0.12597158 0.058017427 0.22403730 0.136694935  0.16569455
#>    66   59 0.14169986 0.047597221 0.22962010 0.099653739  0.20093839
#>    20   22 0.02576702 0.037445920 0.02946219 0.061416008  0.07708611
#>    83   22 0.09361727 0.159676782 0.02946219 0.061416008  0.07132169
#>    74   56 0.07281106 0.366496206 0.01154638 0.101679416 -0.43049307
#>    56   53 0.01154638 0.101679416 0.00000000 0.916430928 -3.37364320
#>    66   76 0.14169986 0.047597221 0.23787853 0.001510489  0.10144937
#>    20   69 0.02576702 0.037445920 0.12597158 0.058017427  0.08878675
#>    20   56 0.02576702 0.037445920 0.01154638 0.101679416 -0.03127252
#>     6   86 0.61712288 0.779287687 0.58410591 0.804542001  0.57455050
#>    43   15 0.84626390 0.628593337 0.68540754 0.652365769  0.76949552
#>    75    3 0.36347308 0.955071926 0.36719013 0.826663324  0.35422837
#>    51   73 0.46802100 0.849171133 0.49674651 0.950103171  0.46265296
#>    35   99 0.52314194 0.520758748 0.48210467 0.471884766  0.54198273
#>    73   18 0.49674651 0.950103171 0.56539786 0.966417381  0.50663474
#>    42   18 0.61014115 0.971043767 0.56539786 0.966417381  0.59612215
#>    84   96 0.23844051 0.667645097 0.33283659 0.552836555  0.28667166
#>    38   75 0.39394805 0.982036406 0.36347308 0.955071926  0.39338527
#>    36   47 0.83455066 0.857745918 0.69041807 0.896394933  0.73785351
#>     4   23 0.16283406 0.619203685 0.20478466 0.617837041  0.18539565
#>     4   10 0.16283406 0.619203685 0.09913948 0.641885370  0.13659724
#>    34   89 1.00000000 0.680471016 0.95869074 0.776378352  0.96152308
#>    60   26 0.82805874 0.396482289 0.87576094 0.450803821  0.81642755
#>    19    9 0.20568461 0.963399359 0.17879399 0.969079531  0.17133669
#>    33   61 0.67555588 0.943695073 0.90419539 1.000000000  0.73761887
#>    21    2 0.53860805 0.264155914 0.31344089 0.259979208  0.42561052
#>    61   38 0.90419539 1.000000000 0.39394805 0.982036406  0.58922109
#>    10   50 0.09913948 0.641885370 0.06933649 0.863553743  0.11193485
#>    24   13 0.34763999 0.183093163 0.35700441 0.177564949  0.39306709
#>    65   13 0.53672626 0.255558464 0.35700441 0.177564949  0.43611389
#>    29    5 0.49175409 0.156311258 0.41586367 0.125379462  0.45782392
#>    95   12 0.45846331 0.080601040 0.46837686 0.089996649  0.44215365
#>    83   17 0.09361727 0.159676782 0.23185851 0.296358543  0.13122043
#>    40   81 0.95517798 0.336311050 0.83458993 0.131957994  0.90166271
#>    48   37 0.45782894 0.001918630 0.66781593 0.052517428  0.56378376
#>    81   97 0.83458993 0.131957994 0.83576524 0.043106610  0.84851248
#>    90   34 0.96440074 0.106119471 1.00000000 0.680471016  2.98749684
#>    94   90 0.89352319 0.032496758 0.96440074 0.106119471  0.92354228
#>    76   85 0.23787853 0.001510489 0.81219968 0.000000000  0.42115206
#>    66   20 0.14169986 0.047597221 0.02576702 0.037445920  0.10144937
#>            my1          mx2           my2 bp1 bp2
#>    0.709959022   0.56627501   0.611130233   0   0
#>    0.758194715   0.53062674   0.709959022   0   0
#>    0.757839440   0.56752767   0.758194715   0   0
#>    0.594263424   0.40838734   0.734506940   0   0
#>    0.781622770   0.35422837   0.890546220   0   0
#>    0.770123863   0.42147853   0.745662102   0   0
#>    0.819573730   0.56752767   0.758194715   0   0
#>    0.715794361   0.57455050   0.757839440   0   0
#>    0.611130233   0.47808225   0.594263424   0   0
#>    0.604079618   0.61169506   0.577413337   0   0
#>    0.745662102   0.40838734   0.734506940   0   0
#>    0.665243637   0.61169506   0.577413337   0   0
#>    0.733235880   0.33085694   0.743580447   0   0
#>    0.896497109   0.38530856   0.905087689   0   0
#>    0.905252586   0.46024641   0.939433378   0   0
#>    0.881253859   0.52326355   0.819573730   0   0
#>    0.865766063   0.64157931   0.845471529   0   0
#>    0.463273432   0.46803100   0.381527871   0   0
#>    0.550817742   0.47808225   0.594263424   0   0
#>    0.625744450   0.28667166   0.611090253   0   0
#>    0.728611561   0.32889601   0.625744450   0   0
#>    0.845471529   0.73785351   0.785215344   0   0
#>    0.766119008   0.65120505   0.715794361   0   0
#>    0.887446179   0.62404498   0.865766063   0   0
#>    0.544408094   0.39821787   0.495299369   0   0
#>    1.061094805   0.52302435   1.450094444   0   0
#>    0.882511954   0.54697663   0.881253859   0   0
#>    0.887949466   0.54907291   0.882511954   0   0
#>    0.891158094   0.32908849   0.889818494   0   0
#>    0.921971550   0.59934079   0.887446179   0   0
#>    0.935867708   0.62404498   0.865766063   0   0
#>    0.968822444   0.46024641   0.939433378   0   0
#>    1.450094444   0.41523041   0.968822444   0   0
#>    0.719212915   0.33085694   0.743580447   0   0
#>    0.951968974   0.38530856   0.905087689   0   0
#>    1.527076631   0.58922109   2.691046877   0   0
#>    0.493386432   0.66098726   0.481960063   0   0
#>    0.381527871   0.46690636   0.344652386   0   0
#>    0.740483180   0.83210591   0.766296639   0   0
#>    0.582599440   0.84433648   0.568379224   0   0
#>    0.495299369   0.31017542   0.465570843   0   0
#>    0.465570843   0.31085126   0.408245318   0   0
#>    0.949716631   0.78731687   0.982238001   0   0
#>    0.604333128   0.18539565   0.667213256   0   0
#>    0.481960063   0.64491250   0.376956321   0   0
#>    0.863500474   0.18012143   0.845363034   0   0
#>    0.847377188   0.33085694   0.743580447   0   0
#>    1.023667767   0.17410726   0.863500474   0   0
#>    0.667605433   0.84433648   0.568379224   0   0
#>    0.959823271   0.92842418   0.867375170   0   0
#>    0.928874914   0.88711811   0.840542913   0   0
#>    0.739135492   0.88071704   0.667605433   0   0
#>    0.551338089   0.13659724   0.646299827   0   0
#>    0.741570970   0.24695311   0.719212915   0   0
#>    0.450695955   0.81642755   0.454801665   0   0
#>    0.486769400   0.76889941   0.450695955   0   0
#>    0.337674599   0.78158403   0.392186662   0   0
#>    0.514806224   0.21625946   0.511945354   0   0
#>    0.646299827   0.08448046   0.580334747   0   0
#>    0.845363034   0.13838734   0.759999859   0   0
#>    1.068427877   0.26592480   1.315075585   0   0
#>    0.555192382   0.95350838   0.625044038   0   0
#>    0.720748257   0.92298235   0.752503274   0   0
#>    0.653105394   0.88071704   0.667605433   0   0
#>    0.392186662   0.85029938   0.279105179   0   0
#>    0.519556705   1.07144148   0.555853175   0   0
#>    0.454801665   0.87840300   0.511334225   0   0
#>    0.869447821   1.76215797   1.090629766   0   0
#>    0.752503274   0.97845446   0.869447821   0   0
#>    0.305684563   0.31085126   0.408245318   0   0
#>    0.580334747   0.04985262   0.606351232   0   0
#>    0.708696894   0.18539565   0.667213256   0   0
#>    0.867283937   0.08536036   0.956463596   0   0
#>    0.344652386   0.42532609   0.299717192   0   0
#>    1.184048530   0.58922109   2.691046877   0   0
#>    0.327057490   0.61495905   0.278981041   0   0
#>    0.247233759   0.57538832   0.208825498   0   0
#>    0.284383769   0.42525371   0.266361778   0   0
#>    0.555853175   1.08889589   0.549815538   0   0
#>    0.511334225   0.91040246   0.472273852   0   0
#>    0.472273852   0.92627900   0.466272969   0   0
#>    1.123245825   4.76746408   1.984199646   0   1
#>    0.404326395   0.89066309   0.364378553   0   0
#>    0.200563968   0.72680683   0.144817434   0   0
#>    1.090629766   0.84246157   0.959823271   0   0
#>    0.435838695   0.21625946   0.511945354   0   0
#>    0.533207199   0.18162073   0.551338089   0   0
#>    0.756443359  -0.04896926   0.780320627   0   0
#>    0.759999859   0.15273777   0.741570970   0   0
#>    0.606351232   0.11193485   0.756443359   0   0
#>    0.351105351   0.21337899   0.354084792   0   0
#>    0.364378553   0.85029938   0.279105179   0   0
#>    0.297034699   0.27806167   0.290305468   0   0
#>    0.450425780   0.10031644   0.550759736   0   0
#>    0.224188437   0.27806167   0.290305468   0   0
#>    0.372320292   1.12310188   0.227758473   0   0
#>    0.278981041   0.74703271   0.200563968   0   0
#>    0.201974634   0.43611389   0.241336452   0   0
#>    0.249348187   0.42525371   0.266361778   0   0
#>    0.241336452   0.42962312   0.200184541   0   0
#>    0.185544538   0.33650320   0.224188437   0   0
#>    0.216332384   0.27806167   0.290305468   0   0
#>    0.364359574   1.08889589   0.549815538   0   0
#>    0.398815805   0.92627900   0.466272969   0   0
#>    0.242110319   0.84881405   0.261320217   0   0
#>    0.051436125   0.51887490   0.109472858   0   0
#>    0.674940559   0.04985262   0.606351232   0   0
#>    0.265866614   0.11446509   0.416315609   0   0
#>    0.416315609   0.17227374   0.435838695   0   0
#>    0.130994545   0.44215365   0.107737608   0   0
#>    0.131857407   0.42962312   0.200184541   0   0
#>    0.092432072   0.44215365   0.107737608   0   0
#>    0.034227620   0.56378376   0.023228502   0   0
#>    0.123513319   0.57538832   0.208825498   0   0
#>    0.259894715   0.21337899   0.354084792   0   0
#>    0.108982196   0.30843744   0.056866870   0   0
#>    0.128825976   0.27620991   0.185544538   0   0
#>    0.230134399   0.84851248   0.087708695   0   0
#>    0.261320217   0.85029938   0.279105179   0   0
#>    0.078074227   0.56378376   0.023228502   0   0
#>    0.041578248   0.50552025   0.040877881   0   0
#>    0.110005315   0.29750104   0.128825976   0   0
#>    0.123042296   0.17135839   0.219298807   0   0
#>    0.227758473   2.98749684   0.269003646   0   0
#>    0.087708695   0.87071874   0.070870174   0   0
#>    0.251421003   0.07132169   0.104159685   0   0
#>    0.086458856   0.72680683   0.144817434   0   0
#>    0.060626686   0.85822113   0.002835674   0   0
#>    0.269003646   2.19749666   0.364359574   0   0
#>    0.220213849   0.92354228   0.074525715   0   0
#>    0.002835674   5.09533345 -10.600586982   0   1
#>   -0.266374335   0.42115206 -39.499378172   0   0
#>   -0.011401958   0.41865300   0.041578248   0   0
#>    0.103939425   0.16569455   0.108960281   0   0
#>    0.108960281   0.29750104   0.128825976   0   0
#>    0.047821185   0.30843744   0.056866870   0   0
#>    0.041804567   0.07910303   0.099079199   0   0
#>    0.104159685  -0.03127252   0.058508847   0   0
#>    0.343439223  -3.37364320   0.461163255   0   0
#>    0.461163255  -0.70677203   0.674940559   0   0
#>   -0.159803029  -0.67498486  -4.742762818   0   1
#>   -0.015189680   0.16771146   0.103939425   0   0
#>    0.781622770   0.43273863   0.770123863   0   0
#>    0.770123863   0.52326355   0.819573730   0   0
#>    0.715794361   0.59006712   0.604079618   0   0
#>    0.611130233   0.59006712   0.604079618   0   0
#>    0.745662102   0.53062674   0.709959022   0   0
#>    0.665243637   0.78704122   0.582599440   0   0
#>    0.577413337   0.64678596   0.493386432   0   0
#>    0.733235880   0.37521380   0.728611561   0   0
#>    0.890546220   0.38530856   0.905087689   0   0
#>    0.896497109   0.43017181   0.781622770   0   0
#>    0.905252586   0.44419854   0.896497109   0   0
#>    0.463273432   0.64678596   0.493386432   0   0
#>    0.550817742   0.42485075   0.544408094   0   0
#>    0.625744450   0.42485075   0.544408094   0   0
#>    0.728611561   0.40838734   0.734506940   0   0
#>    0.766119008   0.78004664   0.740483180   0   0
#>    1.061094805   0.46024641   0.939433378   0   0
#>    0.887949466   0.59934079   0.887446179   0   0
#>    0.891158094   0.28797931   1.023667767   0   0
#>    0.921971550   0.73334404   0.935867708   0   0
#>    0.935867708   0.77741938   0.949716631   0   0
#>    0.611090253   0.24695311   0.719212915   0   0
#>    0.889818494   0.36708655   0.733235880   0   0
#>    0.951968974   0.41523041   0.968822444   0   0
#>    1.527076631   0.52302435   1.450094444   0   0
#>    0.785215344   0.74474437   0.766119008   0   0
#>    0.495299369   0.46803100   0.381527871   0   0
#>    0.604333128   0.23300857   0.514806224   0   0
#>    0.481960063   0.70723718   0.486769400   0   0
#>    0.847377188   0.32435578   0.891158094   0   0
#>    0.568379224   0.86524412   0.555192382   0   0
#>    0.928874914   0.77741938   0.949716631   0   0
#>    0.766296639   0.92298235   0.752503274   0   0
#>    0.739135492   0.78004664   0.740483180   0   0
#>    0.486769400   0.78704122   0.582599440   0   0
#>    0.337674599   0.64491250   0.376956321   0   0
#>    0.840542913   0.83210591   0.766296639   0   0
#>    0.514806224   0.31017542   0.465570843   0   0
#>    0.845363034   0.25153349   0.847377188   0   0
#>    1.068427877   0.28797931   1.023667767   0   0
#>    0.408245318   0.25158578   0.351105351   0   0
#>    0.511945354   0.18417775   0.533207199   0   0
#>    0.376956321   0.60473732   0.327057490   0   0
#>    0.720748257   0.93810180   0.653105394   0   0
#>    0.653105394   0.95350838   0.625044038   0   0
#>    0.392186662   0.76889941   0.450695955   0   0
#>    0.519556705   0.86524412   0.555192382   0   0
#>    0.869447821   0.92842418   0.867375170   0   0
#>    0.305684563   0.28106237   0.297034699   0   0
#>    0.580334747   0.10031644   0.550759736   0   0
#>    0.708696894   0.15273777   0.741570970   0   0
#>    0.867283937   0.17410726   0.863500474   0   0
#>    1.315075585  -0.01296267   5.946112823   0   1
#>    0.344652386   0.60473732   0.327057490   0   0
#>    1.184048530   0.62804932   0.921971550   0   0
#>    0.867375170   0.88711811   0.840542913   0   0
#>    0.247233759   0.61495905   0.278981041   0   0
#>    0.284383769   0.59534087   0.247233759   0   0
#>    0.555853175   0.95350838   0.625044038   0   0
#>    0.511334225   0.87923893   0.519556705   0   0
#>    1.123245825   1.76215797   1.090629766   0   0
#>    0.404326395   0.91040246   0.472273852   0   0
#>    0.200563968   0.79455455   0.242110319   0   0
#>    0.982238001   0.84246157   0.959823271   0   0
#>    0.533207199   0.10031644   0.550759736   0   0
#>    0.299717192   0.41210280   0.305684563   0   0
#>    0.351105351   0.28106237   0.297034699   0   0
#>    0.549815538   0.92627900   0.466272969   0   0
#>    0.364378553   0.90802997   0.372320292   0   0
#>    0.354084792   0.17227374   0.435838695   0   0
#>    0.450425780  -0.43049307   0.343439223   0   0
#>    0.372320292   0.95778167   0.398815805   0   0
#>    0.201974634   0.51887490   0.109472858   0   0
#>    0.249348187   0.33650320   0.224188437   0   0
#>    0.208825498   0.52297983   0.201974634   0   0
#>    0.185544538   0.23191551   0.216332384   0   0
#>    0.216332384   0.17135839   0.219298807   0   0
#>    0.364359574   0.95778167   0.398815805   0   0
#>    0.242110319   0.71240330   0.337674599   0   0
#>    0.051436125   0.56105171   0.034227620   0   0
#>    0.674940559  -0.04896926   0.780320627   0   0
#>    0.780320627   0.08536036   0.956463596   0   0
#>    0.956463596  -1.43017986   6.103219017   0   1
#>    0.265866614  -0.03274405   0.251421003   0   0
#>    0.416315609   0.02003844   0.450425780   0   0
#>    0.130994545   0.51887490   0.109472858   0   0
#>    0.131857407   0.31143291   0.108982196   0   0
#>    0.092432072   0.36904327   0.131857407   0   0
#>    0.123513319   0.56416662   0.051436125   0   0
#>    0.259894715   0.11084934   0.265866614   0   0
#>    0.230134399   0.93479505   0.220213849   0   0
#>    0.144817434   0.62543147   0.123513319   0   0
#>    0.078074227   0.75250864   0.060626686   0   0
#>    0.041578248   0.42606531   0.092432072   0   0
#>    0.110005315   0.31143291   0.108982196   0   0
#>    0.123042296   0.07910303   0.099079199   0   0
#>    0.040877881   0.56105171   0.034227620   0   0
#>    0.086458856   0.75348629   0.078074227   0   0
#>    0.220213849   1.12310188   0.227758473   0   0
#>    0.074525715   0.87071874   0.070870174   0   0
#>    0.002835674   0.87071874   0.070870174   0   0
#>   -0.266374335   0.75250864   0.060626686   0   0
#>  -39.499378172   0.34787807  -0.011401958   0   0
#>   -0.011401958   0.30843744   0.056866870   0   0
#>    0.108960281   0.15439663   0.123042296   0   0
#>    0.047821185   0.16771146   0.103939425   0   0
#>    0.041804567  -0.03127252   0.058508847   0   0
#>    0.104159685   0.07910303   0.099079199   0   0
#>    0.343439223  -0.03274405   0.251421003   0   0
#>    0.461163255  -4.60076483   0.443772912   0   1
#>   -0.159803029   0.20093839   0.047821185   0   0
#>   -0.015189680   0.07708611   0.041804567   0   0
#>    0.058508847 -14.87207456  -3.227094033   0   1
#>    0.757839440   0.64157931   0.845471529   0   0
#>    0.665243637   0.78041576   0.739135492   0   0
#>    0.890546220   0.32908849   0.889818494   0   0
#>    0.905252586   0.54697663   0.881253859   0   0
#>    0.463273432   0.43772046   0.550817742   0   0
#>    1.061094805   0.54907291   0.882511954   0   0
#>    0.887949466   0.53003740   1.527076631   0   0
#>    0.611090253   0.27845336   0.604333128   0   0
#>    0.951968974   0.29034157   1.068427877   0   0
#>    0.785215344   0.77637569   0.928874914   0   0
#>    0.667213256   0.18162073   0.551338089   0   0
#>    0.646299827   0.15881689   0.708696894   0   0
#>    0.720748257   1.89599796   1.123245825   0   0
#>    0.454801665   0.87390695   0.404326395   0   0
#>    0.867283937   0.26592480   1.315075585   0   0
#>    1.184048530   0.78731687   0.982238001   0   0
#>    0.284383769   0.42532609   0.299717192   0   0
#>    2.691046877   0.52030926   4.648454646   0   1
#>    0.756443359   0.13838734   0.759999859   0   0
#>    0.249348187   0.31080715   0.110005315   0   0
#>    0.241336452   0.42525371   0.266361778   0   0
#>    0.130994545   0.42962312   0.200184541   0   0
#>    0.107737608   0.50552025   0.040877881   0   0
#>    0.259894715   0.17135839   0.219298807   0   0
#>    0.230134399   0.84881405   0.261320217   0   0
#>    0.023228502   0.63356691  -0.266374335   0   0
#>    0.087708695   0.75402739   0.086458856   0   0
#>    0.269003646   4.72192758   0.161500754   0   1
#>    0.074525715   7.97287239  -6.711956292   0   1
#>  -39.499378172   0.41657270 -41.240552209   0   1
#>   -0.159803029   0.08878675  -0.015189680   0   0
#> 
#> $x
#>              [,1]        [,2]
#>   [1,] 0.51831685 0.757097527
#>   [2,] 0.31344089 0.259979208
#>   [3,] 0.36719013 0.826663324
#>   [4,] 0.16283406 0.619203685
#>   [5,] 0.41586367 0.125379462
#>   [6,] 0.61712288 0.779287687
#>   [7,] 0.92780640 0.557419866
#>   [8,] 0.28682332 0.685418383
#>   [9,] 0.17879399 0.969079531
#>  [10,] 0.09913948 0.641885370
#>  [11,] 0.67956508 0.913316393
#>  [12,] 0.46837686 0.089996649
#>  [13,] 0.35700441 0.177564949
#>  [14,] 0.91294921 0.938452123
#>  [15,] 0.68540754 0.652365769
#>  [16,] 0.49283865 0.679207890
#>  [17,] 0.23185851 0.296358543
#>  [18,] 0.56539786 0.966417381
#>  [19,] 0.20568461 0.963399359
#>  [20,] 0.02576702 0.037445920
#>  [21,] 0.53860805 0.264155914
#>  [22,] 0.02946219 0.061416008
#>  [23,] 0.20478466 0.617837041
#>  [24,] 0.34763999 0.183093163
#>  [25,] 0.81788525 0.514251722
#>  [26,] 0.87576094 0.450803821
#>  [27,] 0.26251826 0.769107106
#>  [28,] 0.88802002 0.112019023
#>  [29,] 0.49175409 0.156311258
#>  [30,] 0.30828918 0.338024685
#>  [31,] 0.71423987 0.600564108
#>  [32,] 0.44011597 0.971695847
#>  [33,] 0.67555588 0.943695073
#>  [34,] 1.00000000 0.680471016
#>  [35,] 0.52314194 0.520758748
#>  [36,] 0.83455066 0.857745918
#>  [37,] 0.66781593 0.052517428
#>  [38,] 0.39394805 0.982036406
#>  [39,] 0.39163810 0.090341300
#>  [40,] 0.95517798 0.336311050
#>  [41,] 0.73750297 0.376850088
#>  [42,] 0.61014115 0.971043767
#>  [43,] 0.84626390 0.628593337
#>  [44,] 0.66739529 0.053972351
#>  [45,] 0.91163184 0.431537071
#>  [46,] 0.75636274 0.352915334
#>  [47,] 0.69041807 0.896394933
#>  [48,] 0.45782894 0.001918630
#>  [49,] 0.96231590 0.455206657
#>  [50,] 0.06933649 0.863553743
#>  [51,] 0.46802100 0.849171133
#>  [52,] 0.88178782 0.615568756
#>  [53,] 0.00000000 0.916430928
#>  [54,] 0.14096924 0.608837088
#>  [55,] 0.04006796 0.666318977
#>  [56,] 0.01154638 0.101679416
#>  [57,] 0.15367777 0.364552732
#>  [58,] 0.55003801 0.153724802
#>  [59,] 0.22962010 0.099653739
#>  [60,] 0.82805874 0.396482289
#>  [61,] 0.90419539 1.000000000
#>  [62,] 0.93826972 0.502012923
#>  [63,] 0.11426554 0.481253969
#>  [64,] 0.25808914 0.926141993
#>  [65,] 0.53672626 0.255558464
#>  [66,] 0.14169986 0.047597221
#>  [67,] 0.63585357 0.204063228
#>  [68,] 0.69073715 0.296500699
#>  [69,] 0.12597158 0.058017427
#>  [70,] 0.57934396 0.710410992
#>  [71,] 0.38109901 0.409905081
#>  [72,] 0.24058390 0.408248479
#>  [73,] 0.49674651 0.950103171
#>  [74,] 0.07281106 0.366496206
#>  [75,] 0.36347308 0.955071926
#>  [76,] 0.23787853 0.001510489
#>  [77,] 0.85865842 0.887969978
#>  [78,] 0.55653073 0.404554679
#>  [79,] 0.42261136 0.944883583
#>  [80,] 0.95917671 0.342260156
#>  [81,] 0.83458993 0.131957994
#>  [82,] 0.22403730 0.136694935
#>  [83,] 0.09361727 0.159676782
#>  [84,] 0.23844051 0.667645097
#>  [85,] 0.81219968 0.000000000
#>  [86,] 0.58410591 0.804542001
#>  [87,] 0.30381473 0.251470547
#>  [88,] 0.90637839 0.712887409
#>  [89,] 0.95869074 0.776378352
#>  [90,] 0.96440074 0.106119471
#>  [91,] 0.02944372 0.549120584
#>  [92,] 0.91888130 0.795261778
#>  [93,] 0.40146469 0.633799017
#>  [94,] 0.89352319 0.032496758
#>  [95,] 0.45846331 0.080601040
#>  [96,] 0.33283659 0.552836555
#>  [97,] 0.83576524 0.043106610
#>  [98,] 0.32417201 0.816224074
#>  [99,] 0.48210467 0.471884766
#> [100,] 0.40968833 0.856293847
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,16,70), [9,2,11]
#> 2: (86,1,70), [1,3,7]
#> 3: (6,86,70), [2,8,25]
#> 4: (35,16,93), [12,22,9]
#> 5: (51,100,3), [16,6,17]
#> 6: (51,3,1), [11,7,5]
#> 7: (51,1,86), [2,19,6]
#> 8: (15,6,70), [3,10,26]
#> 9: (70,16,35), [4,10,1]
#> 10: (70,35,15), [14,8,9]
#> 11: (1,3,16), [12,1,6]
#> 12: (3,93,16), [4,11,24]
#> 13: (43,15,31), [14,49,64]
#> 14: (31,15,35), [10,45,13]
#> 15: (3,98,8), [33,24,42]
#> 16: (75,3,100), [5,29,42]
#> 17: (51,79,100), [29,5,18]
#> 18: (51,73,79), [38,17,19]
#> 19: (73,51,86), [7,31,18]
#> 20: (11,86,47), [25,36,27]
#> 21: (35,99,78), [46,45,22]
#> 22: (99,35,93), [4,28,21]
#> 23: (93,8,96), [40,28,24]
#> 24: (3,8,93), [23,12,15]
#> 25: (86,6,47), [47,20,3]
#> 26: (36,6,15), [8,48,47]
#> 27: (42,86,11), [20,35,32]
#> 28: (93,96,99), [50,22,23]
#> 29: (75,100,79), [17,43,16]
#> 30: (73,18,32), [39,38,31]
#> 31: (18,73,86), [19,32,30]
#> 32: (42,18,86), [31,27,44]
#> 33: (27,8,98), [15,57,41]
#> 34: (64,98,75), [42,58,57]
#> 35: (33,42,11), [27,36,93]
#> 36: (33,11,47), [20,52,35]
#> 37: (38,79,32), [38,39,43]
#> 38: (73,32,79), [37,18,30]
#> 39: (18,38,32), [37,30,44]
#> 40: (84,96,8), [23,41,53]
#> 41: (84,8,27), [33,66,40]
#> 42: (3,75,98), [34,15,16]
#> 43: (38,75,79), [29,37,74]
#> 44: (18,42,38), [105,39,32]
#> 45: (35,78,31), [55,14,21]
#> 46: (99,71,78), [92,21,50]
#> 47: (36,47,6), [25,26,62]
#> 48: (15,88,36), [63,26,64]
#> 49: (31,25,43), [60,13,68]
#> 50: (99,96,71), [51,46,28]
#> 51: (96,72,71), [75,50,71]
#> 52: (47,77,33), [106,36,62]
#> 53: (96,84,23), [54,71,40]
#> 54: (4,23,84), [53,88,65]
#> 55: (31,78,41), [78,68,45]
#> 56: (19,50,64), [73,58,89]
#> 57: (64,27,98), [33,34,73]
#> 58: (75,19,64), [56,34,74]
#> 59: (88,43,52), [60,80,64]
#> 60: (52,43,25), [49,77,59]
#> 61: (61,77,14), [94,107,106]
#> 62: (47,36,77), [70,52,47]
#> 63: (92,36,88), [48,85,70]
#> 64: (15,43,88), [59,48,13]
#> 65: (23,4,54), [72,109,54]
#> 66: (10,84,27), [41,111,88]
#> 67: (41,60,25), [83,68,81]
#> 68: (31,41,25), [67,49,55]
#> 69: (68,46,41), [81,78,135]
#> 70: (92,77,36), [62,63,94]
#> 71: (96,23,72), [76,51,53]
#> 72: (4,10,54), [87,65,88]
#> 73: (64,50,27), [111,57,56]
#> 74: (75,38,19), [90,58,43]
#> 75: (30,71,72), [51,114,86]
#> 76: (63,72,23), [71,109,108]
#> 77: (25,7,52), [102,60,82]
#> 78: (68,41,78), [55,95,69]
#> 79: (34,89,88), [85,80,101]
#> 80: (34,88,52), [59,102,79]
#> 81: (41,46,60), [91,67,69]
#> 82: (25,62,7), [98,77,99]
#> 83: (60,26,25), [99,67,103]
#> 84: (92,89,14), [107,94,85]
#> 85: (88,89,92), [84,63,79]
#> 86: (2,71,30), [75,118,113]
#> 87: (54,10,91), [112,126,72]
#> 88: (10,4,84), [54,66,72]
#> 89: (19,9,50), [139,56,90]
#> 90: (9,19,38), [74,0,89]
#> 91: (40,60,46), [81,154,117]
#> 92: (78,71,21), [113,95,46]
#> 93: (33,61,42), [105,35,106]
#> 94: (92,14,77), [61,70,84]
#> 95: (78,21,68), [124,78,92]
#> 96: (21,65,67), [128,124,97]
#> 97: (21,2,65), [132,96,113]
#> 98: (7,62,34), [115,102,82]
#> 99: (25,26,62), [100,82,83]
#> 100: (26,45,62), [116,99,103]
#> 101: (89,34,61), [0,107,79]
#> 102: (34,52,7), [77,98,80]
#> 103: (26,60,45), [117,100,83]
#> 104: (68,67,81), [155,135,124]
#> 105: (61,38,42), [44,93,0]
#> 106: (61,33,77), [52,61,93]
#> 107: (89,61,14), [61,84,101]
#> 108: (57,72,63), [76,141,119]
#> 109: (63,23,54), [65,126,76]
#> 110: (10,50,55), [138,112,111]
#> 111: (50,10,27), [66,73,110]
#> 112: (91,10,55), [110,137,87]
#> 113: (2,21,71), [92,86,97]
#> 114: (30,72,17), [119,118,75]
#> 115: (49,34,62), [98,116,133]
#> 116: (49,62,45), [100,134,115]
#> 117: (45,60,40), [91,123,103]
#> 118: (30,17,2), [122,86,114]
#> 119: (57,17,72), [114,108,150]
#> 120: (74,63,91), [126,183,141]
#> 121: (24,2,87), [122,130,127]
#> 122: (87,2,17), [118,131,121]
#> 123: (45,40,80), [162,134,117]
#> 124: (21,67,68), [104,95,96]
#> 125: (58,65,29), [129,147,128]
#> 126: (54,91,63), [120,109,87]
#> 127: (24,13,2), [132,121,160]
#> 128: (58,67,65), [96,125,148]
#> 129: (65,13,29), [143,125,132]
#> 130: (82,24,87), [121,131,152]
#> 131: (82,87,17), [122,164,130]
#> 132: (13,65,2), [97,127,129]
#> 133: (80,34,49), [115,134,169]
#> 134: (80,49,45), [116,123,133]
#> 135: (68,81,46), [154,69,104]
#> 136: (44,58,12), [147,146,148]
#> 137: (53,91,55), [112,138,184]
#> 138: (53,55,50), [110,139,137]
#> 139: (53,50,9), [89,0,138]
#> 140: (83,57,74), [141,165,150]
#> 141: (74,57,63), [108,120,140]
#> 142: (29,5,12), [149,147,143]
#> 143: (5,29,13), [129,144,142]
#> 144: (39,5,13), [143,151,145]
#> 145: (39,95,5), [149,144,158]
#> 146: (12,48,44), [157,136,166]
#> 147: (29,12,58), [136,125,142]
#> 148: (44,67,58), [128,136,155]
#> 149: (95,12,5), [142,145,166]
#> 150: (83,17,57), [119,140,164]
#> 151: (13,59,39), [159,144,160]
#> 152: (59,24,82), [130,177,160]
#> 153: (40,81,28), [163,170,154]
#> 154: (81,40,46), [91,135,153]
#> 155: (44,81,67), [104,148,167]
#> 156: (97,44,37), [157,168,167]
#> 157: (48,37,44), [156,146,173]
#> 158: (39,48,95), [166,145,175]
#> 159: (76,39,59), [151,178,175]
#> 160: (13,24,59), [152,151,127]
#> 161: (69,82,83), [164,181,177]
#> 162: (40,90,80), [169,123,170]
#> 163: (81,97,28), [179,153,167]
#> 164: (17,83,82), [161,131,150]
#> 165: (74,56,83), [182,140,183]
#> 166: (12,95,48), [158,146,149]
#> 167: (97,81,44), [155,156,163]
#> 168: (37,85,97), [172,156,173]
#> 169: (90,34,80), [133,162,0]
#> 170: (40,28,90), [171,162,153]
#> 171: (94,90,28), [170,179,0]
#> 172: (97,85,94), [0,179,168]
#> 173: (37,48,85), [174,168,157]
#> 174: (76,85,48), [173,175,0]
#> 175: (76,48,39), [158,159,174]
#> 176: (66,59,69), [177,186,178]
#> 177: (69,59,82), [152,161,176]
#> 178: (66,76,59), [159,176,185]
#> 179: (94,28,97), [163,172,171]
#> 180: (20,69,22), [181,187,186]
#> 181: (83,22,69), [180,161,182]
#> 182: (83,56,22), [187,181,165]
#> 183: (74,91,56), [184,165,120]
#> 184: (56,91,53), [137,0,183]
#> 185: (66,20,76), [0,178,186]
#> 186: (20,66,69), [176,180,185]
#> 187: (20,22,56), [182,0,180]
#> boundary nodes:  38 9 53 56 20 76 85 94 90 34 61 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.091960879 0.076837110 0.094251382 0.102035119 0.051807956 0.166369119 
#>                                                                         
#> 0.081112140 0.078557278 0.161320863 0.208927097 0.193728299 0.059285036 
#>                                                                         
#> 0.136033278 0.089527348 0.074318664 0.124368270 0.140496393 0.100362241 
#>                                                                         
#> 0.166082680 0.140339452 0.125726696 0.138153147 0.144124835 0.144721527 
#>                                                                         
#> 0.169806222 0.125393044 0.169763718 0.162952846 0.144302810 0.090289156 
#>                                                                         
#> 0.020102818 0.032020444 0.047311770 0.087146649 0.060009485 0.216472381 
#>                                                                         
#> 0.251578752 0.175513292 0.161688622 0.117810665 0.150859762 0.140524876 
#>                                                                         
#> 0.191394420 0.060112887 0.183080565 0.198858873 0.077596346 0.064298881 
#>                                                                         
#> 0.037836334 0.074134591 0.038661065 0.103533773 0.024197860 0.104279802 
#>                                                                         
#> 0.118208034 0.159187061 0.030472338 0.212623967 0.053309639 0.215033399 
#>                                                                         
#> 0.189183676 0.074155266 0.082265920 0.100377370 0.083895151 0.056386260 
#>                                                                         
#> 0.085879170 0.168404852 0.044061042 0.102313555 0.116029327 0.089793690 
#>                                                                         
#> 0.152041557 0.214597099 0.295476497 0.155529580 0.111704887 0.223329131 
#>                                                                         
#> 0.188833044 0.080806679 0.075342034 0.333582516 0.090627266 0.211412642 
#>                                                                         
#> 0.062167279 0.145897672 0.064446926 0.199394583 0.207070024 0.063925318 
#>                                                                         
#> 0.112229633 0.140640957 0.089326039 0.108630648 0.012847577 0.007168101 
#>                                                                         
#> 0.107502994 0.108961023 0.093219582 0.136415519 0.081216497 0.084809378 
#>                                                                         
#> 0.228394665 0.055938566 0.234396330 0.103584847 0.117678961 0.080890061 
#>                                                                         
#> 0.123176715 0.063321236 0.078662171 0.061804828 0.215934355 0.099490035 
#>                                                                         
#> 0.103743310 0.162285412 0.132024313 0.057029272 0.199507400 0.001514510 
#>                                                                         
#> 0.078684967 0.144536608 0.132429419 0.236198462 0.086484002 0.100495399 
#>                                                                         
#> 0.184487961 0.049127545 0.340665712 0.076608213 0.087575985 0.354375937 
#>                                                                         
#> 0.110452683 0.111698692 0.037459550 0.098490096 0.096569218 0.044069471 
#>                                                                         
#> 0.447798965 0.368488564 0.215134019 0.018866886 0.103312462 0.104915306 
#>                                                                         
#> 0.120907942 0.197804551 0.081950803 0.134966595 0.207092525 0.162508331 
#>                                                                         
#> 0.109054838 0.058765922 0.105938228 0.120905735 0.180884027 0.106135465 
#>                                                                         
#> 0.195886157 0.253820162 0.060607437 0.168524988 0.109282735 0.030642093 
#>                                                                         
#> 0.049580115 0.051544020 0.044266661 0.046924594 0.172159786 0.231150451 
#>                                                                         
#> 0.118505798 0.143604875 0.224920285 0.128253242 0.119785973 0.168451160 
#>                                                                         
#> 0.083317821 0.229109111 0.134878434 0.092968090 0.104956774 0.171511718 
#>                                                                         
#> 0.157097336 0.158008064 0.097546838 0.163855566 0.172299231 0.099074864 
#>                                                                         
#> 0.134857050 0.092659438 0.118093810 0.143313170 0.078215322 0.126506926 
#>                                                                         
#> 0.141662770 0.168996912 0.215543853 0.141538108 0.070901613 0.110551371 
#>                                                                         
#> 0.114314578 0.008800980 0.142665698 0.121004994 0.230165994 0.040717678 
#>                                                                         
#> 0.218558740 0.120931144 0.130347766 0.164485224 0.087050070 0.052621722 
#>                                                                         
#> 0.104710367 0.097272778 0.187702930 0.101147819 0.058341285 0.084148928 
#>                                                                         
#> 0.102700035 0.139777965 0.159855055 0.112990117 0.086540907 0.202252524 
#>                                                                         
#> 0.253301065 0.087198314 0.186384456 0.207863349 0.122015645 0.070314451 
#>                                                                         
#> 0.093848055 0.042597546 0.154023642 0.213498055 0.234130529 0.153369330 
#>                                                                         
#> 0.168212771 0.067531338 0.149321532 0.106683768 0.088707363 0.168720201 
#>                                                                         
#> 0.230376262 0.079712455 0.058724353 0.153638368 0.219950790 0.177575112 
#>                                                                         
#> 0.125726034 0.102175583 0.024253234 0.117350134 0.271811136 0.814833323 
#>                                                                         
#> 0.106650470 0.102294384 0.065788820 0.041568024 0.162603498 0.128462390 
#>                                                                         
#> 0.104940132 0.063817896 0.070563180 0.044981826 0.148632501 0.040691608 
#>                                                                         
#> 0.149224490 0.041972855 0.067612553 0.104425439 0.072293352 0.027483994 
#>                                                                         
#> 0.235470312 0.225205888 0.510563446 0.223662884 0.010874449 0.195915625 
#>                                                                         
#> 0.081952006 0.013658548 0.194403047 0.237279687 0.215997164 0.088859157 
#>                                                 
#> 0.575453738 0.102195550 0.574323140 0.116376424 
#> 
#> $alpha
#> [1] 0.2996806
#> 
#> attr(,"class")
#> [1] "scree"