Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2         mx1
#>    24   12 1.000000000 0.614964673 0.863862061 0.673701596  0.93008926
#>    24    1 1.000000000 0.614964673 0.856204348 0.647714596  0.92107218
#>    42    1 0.998138746 0.578104263 0.856204348 0.647714596  0.91600850
#>    42   70 0.998138746 0.578104263 0.866296507 0.523671451  0.93224049
#>    67   95 0.824808169 0.515060712 0.812819018 0.488575293  0.85234460
#>    70    1 0.866296507 0.523671451 0.856204348 0.647714596  0.83227571
#>     1   12 0.856204348 0.647714596 0.863862061 0.673701596  0.78401929
#>    64   81 0.879486420 0.807559690 0.814291836 0.786154030  0.86509813
#>    12   77 0.863862061 0.673701596 0.704057454 0.728030573  0.78206092
#>    10   56 0.897230472 0.379940197 0.985756339 0.503690316  0.91694348
#>    67   19 0.824808169 0.515060712 0.711154980 0.649134982  0.75301740
#>    26   54 0.921654213 0.838289577 0.801568218 0.844360184  0.86326859
#>    71   26 0.991743643 0.845216993 0.921654213 0.838289577  0.96765874
#>    64   12 0.879486420 0.807559690 0.863862061 0.673701596  0.96816090
#>    67    1 0.824808169 0.515060712 0.856204348 0.647714596  0.78315612
#>    54   77 0.801568218 0.844360184 0.704057454 0.728030573  0.73671428
#>    54   81 0.801568218 0.844360184 0.814291836 0.786154030  0.83864331
#>    90   54 0.929314354 0.929971860 0.801568218 0.844360184  0.86403469
#>    81   77 0.814291836 0.786154030 0.704057454 0.728030573  0.78586548
#>    14   95 0.812188327 0.366939986 0.812819018 0.488575293  0.78091910
#>    95   10 0.812819018 0.488575293 0.897230472 0.379940197  0.87601691
#>    38   92 0.600983660 0.593434858 0.730132237 0.475589803  0.68083082
#>    92   14 0.730132237 0.475589803 0.812188327 0.366939986  0.77356216
#>    71   90 0.991743643 0.845216993 0.929314354 0.929971860  0.95273552
#>    14   36 0.812188327 0.366939986 0.825383919 0.245310964  0.86424571
#>    14   10 0.812188327 0.366939986 0.897230472 0.379940197  0.84643289
#>    38   29 0.600983660 0.593434858 0.590915138 0.505473225  0.67131974
#>    19   38 0.711154980 0.649134982 0.600983660 0.593434858  0.68743736
#>    82   77 0.561208530 0.794255566 0.704057454 0.728030573  0.67462960
#>    34   54 0.826888539 0.994089949 0.801568218 0.844360184  0.84775596
#>     4   34 0.900287444 0.987738260 0.826888539 0.994089949  0.85835944
#>    79   73 0.936659891 0.275889119 0.989337460 0.247099379  0.99150838
#>    79   43 0.936659891 0.275889119 0.945335040 0.361648738  0.90296332
#>    92   29 0.730132237 0.475589803 0.590915138 0.505473225  0.64995992
#>    49   14 0.696692710 0.307109794 0.812188327 0.366939986  0.72777534
#>    19   77 0.711154980 0.649134982 0.704057454 0.728030573  0.63407092
#>    38   21 0.600983660 0.593434858 0.552523184 0.701199997  0.54875717
#>    34   61 0.826888539 0.994089949 0.538575536 0.986141840  0.68407959
#>    37   21 0.563793727 0.759036924 0.552523184 0.701199997  0.62795348
#>    24   42 1.000000000 0.614964673 0.998138746 0.578104263  8.69928994
#>    42   56 0.998138746 0.578104263 0.985756339 0.503690316  1.90925245
#>    73   56 0.989337460 0.247099379 0.985756339 0.503690316  1.15842980
#>    82   60 0.561208530 0.794255566 0.516375588 0.755932748  0.53894348
#>    37   77 0.563793727 0.759036924 0.704057454 0.728030573  0.64254418
#>    36   79 0.825383919 0.245310964 0.936659891 0.275889119  0.86765665
#>    57   73 0.826826039 0.231932038 0.989337460 0.247099379  0.91433996
#>    36   57 0.825383919 0.245310964 0.826826039 0.231932038  0.88530790
#>    75   57 0.768027434 0.243744818 0.826826039 0.231932038  0.79695276
#>    53   21 0.496941906 0.592919999 0.552523184 0.701199997  0.50796344
#>    60   21 0.516375588 0.755932748 0.552523184 0.701199997  0.54166577
#>    49   91 0.696692710 0.307109794 0.576851216 0.393807857  0.62743050
#>    49   92 0.696692710 0.307109794 0.730132237 0.475589803  0.67226267
#>    65   91 0.450367574 0.512771153 0.576851216 0.393807857  0.51798374
#>     9   75 0.766233971 0.241700603 0.768027434 0.243744818  0.73126224
#>    49   83 0.696692710 0.307109794 0.653550937 0.266543761  0.70496947
#>    55   47 0.738052585 0.214087621 0.751437407 0.112594683  0.79315018
#>    57    9 0.826826039 0.231932038 0.766233971 0.241700603  0.78902342
#>     9   49 0.766233971 0.241700603 0.696692710 0.307109794  0.71927120
#>    82   11 0.561208530 0.794255566 0.533048291 0.900365551  0.49616549
#>    55   83 0.738052585 0.214087621 0.653550937 0.266543761  0.70223496
#>    83   91 0.653550937 0.266543761 0.576851216 0.393807857  0.52633634
#>    53   65 0.496941906 0.592919999 0.450367574 0.512771153  0.52146897
#>    80   48 0.444903918 0.646598461 0.393830872 0.583006520  0.44476726
#>    80   60 0.444903918 0.646598461 0.516375588 0.755932748  0.46799579
#>    11   54 0.533048291 0.900365551 0.801568218 0.844360184  0.67997508
#>    20   83 0.651142951 0.247297607 0.653550937 0.266543761  0.52972390
#>    98   61 0.530034166 0.919813450 0.538575536 0.986141840  0.52361951
#>    54   61 0.801568218 0.844360184 0.538575536 0.986141840  0.68027680
#>    60   96 0.516375588 0.755932748 0.428144015 0.799634479  0.45851021
#>    17   96 0.413740851 0.883069814 0.428144015 0.799634479  0.47923711
#>    48   65 0.393830872 0.583006520 0.450367574 0.512771153  0.44728975
#>    50   80 0.401407006 0.637263008 0.444903918 0.646598461  0.41431745
#>    50   48 0.401407006 0.637263008 0.393830872 0.583006520  0.43097940
#>    63   50 0.311159987 0.659497444 0.401407006 0.637263008  0.36518516
#>    98   41 0.530034166 0.919813450 0.494322153 0.973740172  0.46701363
#>    27   73 0.823080740 0.000000000 0.989337460 0.247099379  0.91975564
#>    55   20 0.738052585 0.214087621 0.651142951 0.247297607  0.66483436
#>    86   25 0.379008711 0.270251878 0.504771541 0.150841024  0.50939676
#>    69   59 0.366965622 0.800600269 0.404040790 0.729618022  0.39709777
#>    17   69 0.413740851 0.883069814 0.366965622 0.800600269  0.39814364
#>    98   17 0.530034166 0.919813450 0.413740851 0.883069814  0.47206749
#>    69   97 0.366965622 0.800600269 0.285352517 0.708826039  0.34075275
#>    65   62 0.450367574 0.512771153 0.355124791 0.482758260  0.43451591
#>    48   63 0.393830872 0.583006520 0.311159987 0.659497444  0.34115025
#>    66   97 0.203288530 0.771398064 0.285352517 0.708826039  0.28402374
#>    33   87 0.337579316 0.910594189 0.336919569 0.876146952  0.37415371
#>    20   25 0.651142951 0.247297607 0.504771541 0.150841024  0.60125413
#>    62   23 0.355124791 0.482758260 0.282460565 0.611423391  0.28795455
#>    62   48 0.355124791 0.482758260 0.393830872 0.583006520  0.32646126
#>    97   23 0.285352517 0.708826039 0.282460565 0.611423391  0.25400844
#>    97   63 0.285352517 0.708826039 0.311159987 0.659497444  0.34642397
#>    47   27 0.751437407 0.112594683 0.823080740 0.000000000  0.87964635
#>    99   27 0.615157588 0.067376796 0.823080740 0.000000000  0.71064946
#>    66   69 0.203288530 0.771398064 0.366965622 0.800600269  0.28365506
#>    41    2 0.494322153 0.973740172 0.356278740 0.996945982  0.42070746
#>    47   20 0.751437407 0.112594683 0.651142951 0.247297607  0.66292617
#>     3   84 0.317978907 0.406203454 0.343676749 0.303661809  0.44525270
#>    84   86 0.343676749 0.303661809 0.379008711 0.270251878  0.44850981
#>    33    2 0.337579316 0.910594189 0.356278740 0.996945982  0.28028361
#>    41   61 0.494322153 0.973740172 0.538575536 0.986141840  0.45890836
#>     5   25 0.504620615 0.104362036 0.504771541 0.150841024  0.39266250
#>    84   91 0.343676749 0.303661809 0.576851216 0.393807857  0.44715740
#>     3   13 0.317978907 0.406203454 0.265788048 0.450797393  0.29356068
#>    99   25 0.615157588 0.067376796 0.504771541 0.150841024  0.57377745
#>     3   62 0.317978907 0.406203454 0.355124791 0.482758260  0.31462745
#>    66   39 0.203288530 0.771398064 0.148220232 0.628454355  0.20506883
#>    52    3 0.311902469 0.352628564 0.317978907 0.406203454  0.29179608
#>     3   16 0.317978907 0.406203454 0.265942454 0.409920083  0.29047263
#>    31    7 0.240140310 0.274802330 0.328638970 0.222795290  0.29713394
#>    52   31 0.311902469 0.352628564 0.240140310 0.274802330  0.28821915
#>    16   13 0.265942454 0.409920083 0.265788048 0.450797393  0.19584866
#>    13   44 0.265788048 0.450797393 0.130066638 0.581221536  0.17669340
#>    97   39 0.285352517 0.708826039 0.148220232 0.628454355  0.21462904
#>    69   87 0.366965622 0.800600269 0.336919569 0.876146952  0.28264337
#>    33   18 0.337579316 0.910594189 0.200015446 0.980791459  0.24301887
#>    99    5 0.615157588 0.067376796 0.504620615 0.104362036  0.54714761
#>     5    7 0.504620615 0.104362036 0.328638970 0.222795290  0.36844415
#>   100    7 0.262894965 0.005049415 0.328638970 0.222795290  0.26476207
#>    52   35 0.311902469 0.352628564 0.257237259 0.390704196  0.25710358
#>     6   39 0.156850045 0.617585293 0.148220232 0.628454355  0.13273675
#>    76    5 0.493759751 0.001326674 0.504620615 0.104362036  0.37933180
#>    45   74 0.182878867 0.325480072 0.217868887 0.288879394  0.23537002
#>    35   13 0.257237259 0.390704196 0.265788048 0.450797393  0.17576650
#>    44   39 0.130066638 0.581221536 0.148220232 0.628454355  0.12833457
#>    28   39 0.105434644 0.624797508 0.148220232 0.628454355  0.12292126
#>    18   68 0.200015446 0.980791459 0.200190942 0.790413146  0.23572672
#>     6   44 0.156850045 0.617585293 0.130066638 0.581221536  0.21628325
#>    18   94 0.200015446 0.980791459 0.200563916 1.000000000  0.19266047
#>    18    2 0.200015446 0.980791459 0.356278740 0.996945982  0.27821927
#>     2   34 0.356278740 0.996945982 0.826888539 0.994089949  0.73193025
#>    44   28 0.130066638 0.581221536 0.105434644 0.624797508  0.08396378
#>    18   30 0.200015446 0.980791459 0.056047397 0.931408513  0.15222820
#>    66   68 0.203288530 0.771398064 0.200190942 0.790413146  0.14188660
#>    18   88 0.200015446 0.980791459 0.182671139 0.997739767  0.13606085
#>    72   45 0.086329120 0.453526014 0.182878867 0.325480072  0.16945517
#>    28   46 0.105434644 0.624797508 0.048750774 0.627436835  0.07572578
#>    22   46 0.107216021 0.720477961 0.048750774 0.627436835  0.07928225
#>    93   46 0.033788501 0.916498416 0.048750774 0.627436835 -0.06891321
#>    40   74 0.077582088 0.329650209 0.217868887 0.288879394  0.12663118
#>    15   40 0.083588733 0.366240077 0.077582088 0.329650209  0.13071171
#>    72   32 0.086329120 0.453526014 0.044098795 0.528860143 -0.02245144
#>    32   44 0.044098795 0.528860143 0.130066638 0.581221536  0.08390127
#>    15   72 0.083588733 0.366240077 0.086329120 0.453526014  0.11593530
#>    15   85 0.083588733 0.366240077 0.069331256 0.382545195  0.05164699
#>     8   40 0.009053839 0.281170657 0.077582088 0.329650209  0.01401878
#>    40   78 0.077582088 0.329650209 0.146697137 0.129878576  0.12601593
#>    30   93 0.056047397 0.931408513 0.033788501 0.916498416  0.10380053
#>    93   22 0.033788501 0.916498416 0.107216021 0.720477961  0.09841761
#>    74   78 0.217868887 0.288879394 0.146697137 0.129878576  0.18300068
#>    93   89 0.033788501 0.916498416 0.037847357 0.581465316 -0.53682851
#>     8   51 0.009053839 0.281170657 0.000000000 0.236858382  0.08824046
#>    32    8 0.044098795 0.528860143 0.009053839 0.281170657 -0.11832297
#>    32   89 0.044098795 0.528860143 0.037847357 0.581465316 -0.54362173
#>   100   78 0.262894965 0.005049415 0.146697137 0.129878576 -0.46765999
#>    51   78 0.000000000 0.236858382 0.146697137 0.129878576  0.10551056
#>    93   51 0.033788501 0.916498416 0.000000000 0.236858382 -2.76612344
#>    56   70 0.985756339 0.503690316 0.866296507 0.523671451  0.93224049
#>    70   95 0.866296507 0.523671451 0.812819018 0.488575293  0.85234460
#>    19   12 0.711154980 0.649134982 0.863862061 0.673701596  0.78401929
#>    12   81 0.863862061 0.673701596 0.814291836 0.786154030  0.86509813
#>    71   24 0.991743643 0.845216993 1.000000000 0.614964673  0.96850798
#>    58   19 0.737271985 0.480755378 0.711154980 0.649134982  0.75301740
#>    64   54 0.879486420 0.807559690 0.801568218 0.844360184  0.86326859
#>    64   26 0.879486420 0.807559690 0.921654213 0.838289577  0.96765874
#>    71   12 0.991743643 0.845216993 0.863862061 0.673701596  0.96816090
#>    19    1 0.711154980 0.649134982 0.856204348 0.647714596  0.78315612
#>    58   95 0.737271985 0.480755378 0.812819018 0.488575293  0.78091910
#>    58   67 0.737271985 0.480755378 0.824808169 0.515060712  0.77103112
#>    70   10 0.866296507 0.523671451 0.897230472 0.379940197  0.87601691
#>    58   92 0.737271985 0.480755378 0.730132237 0.475589803  0.68083082
#>    58   14 0.737271985 0.480755378 0.812188327 0.366939986  0.77356216
#>    43   10 0.945335040 0.361648738 0.897230472 0.379940197  0.94683499
#>    26   90 0.921654213 0.838289577 0.929314354 0.929971860  0.95273552
#>    10   36 0.897230472 0.379940197 0.825383919 0.245310964  0.86424571
#>    58   38 0.737271985 0.480755378 0.600983660 0.593434858  0.68743736
#>    21   19 0.552523184 0.701199997 0.711154980 0.649134982  0.63065361
#>    90   34 0.929314354 0.929971860 0.826888539 0.994089949  0.85835944
#>     4   71 0.900287444 0.987738260 0.991743643 0.845216993  1.36308758
#>    43   73 0.945335040 0.361648738 0.989337460 0.247099379  0.99150838
#>    91   29 0.576851216 0.393807857 0.590915138 0.505473225  0.64995992
#>    21   77 0.552523184 0.701199997 0.704057454 0.728030573  0.63407092
#>    73   42 0.989337460 0.247099379 0.998138746 0.578104263  8.69928994
#>    43   56 0.945335040 0.361648738 0.985756339 0.503690316  1.15842980
#>    37   60 0.563793727 0.759036924 0.516375588 0.755932748  0.53894348
#>    75   49 0.768027434 0.243744818 0.696692710 0.307109794  0.76646211
#>    10   79 0.897230472 0.379940197 0.936659891 0.275889119  0.86765665
#>    79   57 0.936659891 0.275889119 0.826826039 0.231932038  0.88530790
#>    75   14 0.768027434 0.243744818 0.812188327 0.366939986  0.79509391
#>    80   21 0.444903918 0.646598461 0.552523184 0.701199997  0.50796344
#>    91   92 0.576851216 0.393807857 0.730132237 0.475589803  0.67226267
#>    53   38 0.496941906 0.592919999 0.600983660 0.593434858  0.54915264
#>    57   47 0.826826039 0.231932038 0.751437407 0.112594683  0.79315018
#>    55    9 0.738052585 0.214087621 0.766233971 0.241700603  0.78902342
#>    55   49 0.738052585 0.214087621 0.696692710 0.307109794  0.71927120
#>    96   82 0.428144015 0.799634479 0.561208530 0.794255566  0.49582492
#>    96   11 0.428144015 0.799634479 0.533048291 0.900365551  0.49616549
#>    54   82 0.801568218 0.844360184 0.561208530 0.794255566  0.66881889
#>    25   91 0.504771541 0.150841024 0.576851216 0.393807857  0.52633634
#>    29   65 0.590915138 0.505473225 0.450367574 0.512771153  0.52146897
#>    53   48 0.496941906 0.592919999 0.393830872 0.583006520  0.44476726
#>    59   60 0.404040790 0.729618022 0.516375588 0.755932748  0.46799579
#>    98   54 0.530034166 0.919813450 0.801568218 0.844360184  0.67997508
#>    25   83 0.504771541 0.150841024 0.653550937 0.266543761  0.52972390
#>    59   96 0.404040790 0.729618022 0.428144015 0.799634479  0.45851021
#>    59   80 0.404040790 0.729618022 0.444903918 0.646598461  0.41431745
#>    59   50 0.404040790 0.729618022 0.401407006 0.637263008  0.36518516
#>    17   41 0.413740851 0.883069814 0.494322153 0.973740172  0.46701363
#>    96   69 0.428144015 0.799634479 0.366965622 0.800600269  0.39814364
#>    11   17 0.533048291 0.900365551 0.413740851 0.883069814  0.47206749
#>    59   97 0.404040790 0.729618022 0.285352517 0.708826039  0.34075275
#>    23   63 0.282460565 0.611423391 0.311159987 0.659497444  0.34115025
#>    87   17 0.336919569 0.876146952 0.413740851 0.883069814  0.37810834
#>    13   23 0.265788048 0.450797393 0.282460565 0.611423391  0.28795455
#>    23   48 0.282460565 0.611423391 0.393830872 0.583006520  0.32646126
#>    59   63 0.404040790 0.729618022 0.311159987 0.659497444  0.34642397
#>    57   27 0.826826039 0.231932038 0.823080740 0.000000000  0.87964635
#>    68   69 0.200190942 0.790413146 0.366965622 0.800600269  0.28365506
#>    17    2 0.413740851 0.883069814 0.356278740 0.996945982  0.42070746
#>    99   20 0.615157588 0.067376796 0.651142951 0.247297607  0.66292617
#>    65   84 0.450367574 0.512771153 0.343676749 0.303661809  0.44525270
#>    91   86 0.576851216 0.393807857 0.379008711 0.270251878  0.44850981
#>    17   33 0.413740851 0.883069814 0.337579316 0.910594189  0.39265843
#>     2   61 0.356278740 0.996945982 0.538575536 0.986141840  0.45890836
#>     7   25 0.328638970 0.222795290 0.504771541 0.150841024  0.39266250
#>     7   86 0.328638970 0.222795290 0.379008711 0.270251878  0.41470025
#>    13   62 0.265788048 0.450797393 0.355124791 0.482758260  0.31462745
#>    35    3 0.257237259 0.390704196 0.317978907 0.406203454  0.29179608
#>    35   16 0.257237259 0.390704196 0.265942454 0.409920083  0.29047263
#>    84    7 0.343676749 0.303661809 0.328638970 0.222795290  0.29713394
#>    84   31 0.343676749 0.303661809 0.240140310 0.274802330  0.28821915
#>    72   44 0.086329120 0.453526014 0.130066638 0.581221536  0.17669340
#>    68   87 0.200190942 0.790413146 0.336919569 0.876146952  0.28264337
#>    22   66 0.107216021 0.720477961 0.203288530 0.771398064  0.18063612
#>    87   18 0.336919569 0.876146952 0.200015446 0.980791459  0.24301887
#>    76   99 0.493759751 0.001326674 0.615157588 0.067376796  0.65779086
#>    74   52 0.217868887 0.288879394 0.311902469 0.352628564  0.25904543
#>    78    7 0.146697137 0.129878576 0.328638970 0.222795290  0.26476207
#>    74   35 0.217868887 0.288879394 0.257237259 0.390704196  0.25710358
#>     6   97 0.156850045 0.617585293 0.285352517 0.708826039  0.22198343
#>   100    5 0.262894965 0.005049415 0.504620615 0.104362036  0.37933180
#>    72   13 0.086329120 0.453526014 0.265788048 0.450797393  0.17576650
#>    22   39 0.107216021 0.720477961 0.148220232 0.628454355  0.12292126
#>    23   44 0.282460565 0.611423391 0.130066638 0.581221536  0.21628325
#>    88   94 0.182671139 0.997739767 0.200563916 1.000000000  0.19266047
#>    94    2 0.200563916 1.000000000 0.356278740 0.996945982  0.27821927
#>    94   34 0.200563916 1.000000000 0.826888539 0.994089949  0.73193025
#>    89   28 0.037847357 0.581465316 0.105434644 0.624797508  0.08396378
#>    68   30 0.200190942 0.790413146 0.056047397 0.931408513  0.15222820
#>    22   68 0.107216021 0.720477961 0.200190942 0.790413146  0.14188660
#>    30   88 0.056047397 0.931408513 0.182671139 0.997739767  0.13606085
#>    35   45 0.257237259 0.390704196 0.182878867 0.325480072  0.16945517
#>    89   46 0.037847357 0.581465316 0.048750774 0.627436835  0.07572578
#>    45   40 0.182878867 0.325480072 0.077582088 0.329650209  0.13071171
#>    85   32 0.069331256 0.382545195 0.044098795 0.528860143 -0.02245144
#>    89   44 0.037847357 0.581465316 0.130066638 0.581221536  0.08390127
#>    85   72 0.069331256 0.382545195 0.086329120 0.453526014  0.11593530
#>    78   31 0.146697137 0.129878576 0.240140310 0.274802330  0.23959240
#>    40   85 0.077582088 0.329650209 0.069331256 0.382545195  0.05164699
#>    68   93 0.200190942 0.790413146 0.033788501 0.916498416  0.10380053
#>    40   51 0.077582088 0.329650209 0.000000000 0.236858382  0.08824046
#>    85    8 0.069331256 0.382545195 0.009053839 0.281170657 -0.11832297
#>     8   89 0.009053839 0.281170657 0.037847357 0.581465316 -0.54362173
#>    89   51 0.037847357 0.581465316 0.000000000 0.236858382 -2.76612344
#>    70   67 0.866296507 0.523671451 0.824808169 0.515060712  0.85234460
#>    64   71 0.879486420 0.807559690 0.991743643 0.845216993  0.96765874
#>    90    4 0.929314354 0.929971860 0.900287444 0.987738260  0.85835944
#>    73   24 0.989337460 0.247099379 1.000000000 0.614964673  8.69928994
#>    37   82 0.563793727 0.759036924 0.561208530 0.794255566  0.53894348
#>    36   75 0.825383919 0.245310964 0.768027434 0.243744818  0.79695276
#>    80   53 0.444903918 0.646598461 0.496941906 0.592919999  0.50796344
#>    53   29 0.496941906 0.592919999 0.590915138 0.505473225  0.54915264
#>    57   55 0.826826039 0.231932038 0.738052585 0.214087621  0.79315018
#>    98   11 0.530034166 0.919813450 0.533048291 0.900365551  0.67997508
#>     3   65 0.317978907 0.406203454 0.450367574 0.512771153  0.43451591
#>    47   99 0.751437407 0.112594683 0.615157588 0.067376796  0.71064946
#>    52   84 0.311902469 0.352628564 0.343676749 0.303661809  0.39313470
#>    76   27 0.493759751 0.001326674 0.823080740 0.000000000  0.65779086
#>    74   31 0.217868887 0.288879394 0.240140310 0.274802330  0.25904543
#>     6   23 0.156850045 0.617585293 0.282460565 0.611423391  0.22198343
#>   100   76 0.262894965 0.005049415 0.493759751 0.001326674  0.37933180
#>    72   35 0.086329120 0.453526014 0.257237259 0.390704196  0.17576650
#>    22   28 0.107216021 0.720477961 0.105434644 0.624797508  0.12292126
#>    45   15 0.182878867 0.325480072 0.083588733 0.366240077  0.13071171
#>    51  100 0.000000000 0.236858382 0.262894965 0.005049415 -0.46765999
#>           my1          mx2          my2 bp1 bp2
#>   0.640064348   0.96850798  0.729109624   0   0
#>   0.600472911   0.93008926  0.640064348   0   0
#>   0.590148144   0.92107218  0.600472911   0   0
#>   0.550832484   0.91600850  0.590148144   0   0
#>   0.486639524   0.77103112  0.523447689   0   0
#>   0.583335639   0.91600850  0.590148144   0   0
#>   0.683107476   0.93008926  0.640064348   0   0
#>   0.741398228   0.83864331  0.821970900   0   0
#>   0.695280796   0.78586548  0.706471635   0   0
#>   0.459377291   0.94683499  0.437994099   0   0
#>   0.569412895   0.78315612  0.594961136   0   0
#>   0.874110353   0.86403469  0.889264890   0   0
#>   0.730865377   0.95273552  0.881853853   0   0
#>   0.729368401   0.86509813  0.741398228   0   0
#>   0.594961136   0.83227571  0.583335639   0   0
#>   0.799689647   0.67462960  0.851730775   0   0
#>   0.821970900   0.73671428  0.799689647   0   0
#>   0.889264890   0.84775596  0.913555325   0   0
#>   0.706471635   0.73671428  0.799689647   0   0
#>   0.427921409   0.84643289  0.427581713   0   0
#>   0.450569040   0.84643289  0.427581713   0   0
#>   0.551250157   0.67131974  0.540826782   0   0
#>   0.423078877   0.72777534  0.388499090   0   0
#>   0.881853853   1.36308758  1.184113637   0   0
#>   0.311057408   0.79509391  0.303555096   0   0
#>   0.427581713   0.86424571  0.311057408   0   0
#>   0.540826782   0.54915264  0.554810629   0   0
#>   0.559240914   0.63065361  0.671555587   0   0
#>   0.851730775   0.64254418  0.782521742   0   0
#>   0.913555325   0.68407959  0.941234111   0   0
#>   0.930493915   2.02858598 14.453407991   0   1
#>   0.313659418   0.91433996  0.172461737   0   0
#>   0.322616332   0.99150838  0.313659418   0   0
#>   0.441318372   0.67131974  0.540826782   0   0
#>   0.388499090   0.76646211  0.313818534   0   0
#>   0.681967470   0.78206092  0.695280796   0   0
#>   0.634727903   0.63065361  0.671555587   0   0
#>   0.941234111   0.60800205  3.700902203   0   0
#>   0.716517672   0.54166577  0.733332350   0   0
#>   0.207714326   0.92107218  0.600472911   0   0
#>   0.388258666   0.93224049  0.550832484   0   0
#>   0.377779781   1.90925245  0.388258666   0   0
#>   0.774917014   0.49582492  0.825360383   0   0
#>   0.782521742   0.62795348  0.716517672   0   0
#>   0.309237117   0.88530790  0.245003009   0   0
#>   0.172461737   0.91975564  0.114435128   0   0
#>   0.245003009   0.79695276  0.235479171   0   0
#>   0.235479171   0.79372119  0.219393927   0   0
#>   0.655667758   0.54875717  0.634727903   0   0
#>   0.733332350   0.48690787  0.697168144   0   0
#>   0.337546258   0.67226267  0.399517118   0   0
#>   0.399517118   0.72777534  0.388499090   0   0
#>   0.457940379   0.44715740  0.382636704   0   0
#>   0.274191392   0.79372119  0.219393927   0   0
#>   0.255083958   0.62743050  0.337546258   0   0
#>   0.169724795   0.66483436  0.152802591   0   0
#>   0.190254801   0.79372119  0.219393927   0   0
#>   0.261442850   0.73126224  0.274191392   0   0
#>   0.833785649   0.66881889  0.879605662   0   0
#>   0.250678935   0.70496947  0.255083958   0   0
#>   0.276618673   0.62743050  0.337546258   0   0
#>   0.525060824   0.44728975  0.568166218   0   0
#>   0.594402920   0.43097940  0.605476458   0   0
#>   0.709530947   0.48690787  0.697168144   0   0
#>   0.933094468   0.66881889  0.879605662   0   0
#>   0.272262690   0.70223496  0.250678935   0   0
#>   0.954353639   0.68027680  0.934180258   0   0
#>   0.934180258   0.68407959  0.941234111   0   0
#>   0.750023885   0.49582492  0.825360383   0   0
#>   0.851415362   0.39814364  0.837416467   0   0
#>   0.568166218   0.39406277  0.525320599   0   0
#>   0.683109889   0.43097940  0.605476458   0   0
#>   0.605476458   0.34854897  0.616986676   0   0
#>   0.684511046   0.34854897  0.616986676   0   0
#>   0.916867406   0.52361951  0.954353639   0   0
#>   0.114435128   3.70556229 -1.759948695   0   1
#>   0.152802591   0.70223496  0.250678935   0   0
#>   0.281644049   0.41470025  0.181910233   0   0
#>   0.771165170   0.34075275  0.741735265   0   0
#>   0.837416467   0.37810834  0.848780122   0   0
#>   0.900871985   0.46701363  0.916867406   0   0
#>   0.741735265   0.28402374  0.792183313   0   0
#>   0.396946783   0.39406277  0.525320599   0   0
#>   0.608990182   0.34854897  0.616986676   0   0
#>   0.792183313   0.20506883  0.688632973   0   0
#>   0.892663766   0.24301887  0.895175309   0   0
#>   0.163716642   0.52972390  0.272262690   0   0
#>   0.529674852   0.32646126  0.551421695   0   0
#>   0.551421695   0.39406277  0.525320599   0   0
#>   0.661012410   0.22198343  0.661963255   0   0
#>   0.709361870   0.25400844  0.661012410   0   0
#>   0.115082823   0.71064946  0.007551096   0   0
#>   0.007551096   0.65779086 -0.155569253   0   0
#>   0.794249746   0.28402374  0.792183313   0   0
#>   0.958021004   0.45890836  1.185265066   0   0
#>   0.151381823   0.66483436  0.152802591   0   0
#>   0.383608514   0.39313470  0.370547283   0   0
#>   0.379138534   0.33849438  0.262794055   0   0
#>   0.968202107   0.39265843  0.943867413   0   0
#>   1.185265066   0.52361951  0.954353639   0   0
#>   0.127965323   0.57377745  0.127377211   0   0
#>   0.382636704   0.44850981  0.379138534   0   0
#>   0.430463352   0.31462745  0.455119003   0   0
#>   0.127377211   0.60125413  0.163716642   0   0
#>   0.455119003   0.43451591  0.396946783   0   0
#>   0.688632973   0.18063612  0.698045547   0   0
#>   0.382041059   0.39313470  0.370547283   0   0
#>   0.387227643   0.29356068  0.430463352   0   0
#>   0.270485361   0.23959240  0.172568833   0   0
#>   0.302468115   0.25904543  0.329368674   0   0
#>   0.430094264   0.29356068  0.430463352   0   0
#>   0.493913092   0.21803368  0.536932438   0   0
#>   0.672321105   0.20506883  0.688632973   0   0
#>   0.810812265   0.37810834  0.848780122   0   0
#>   0.895175309   0.28028361  0.968202107   0   0
#>   0.047789206   0.57377745  0.127377211   0   0
#>   0.091978934   0.39266250  0.127965323   0   0
#>   0.123283659   0.36844415  0.091978934   0   0
#>   0.332233012   0.29179608  0.382041059   0   0
#>   0.607300309   0.21462904  0.672321105   0   0
#>   0.065478519   0.54714761  0.047789206   0   0
#>   0.340635826   0.12663118  0.236682176   0   0
#>   0.432951797   0.19584866  0.430094264   0   0
#>   0.608992254   0.13273675  0.607300309   0   0
#>   0.672328752   0.12833457  0.608992254   0   0
#>   0.885635141   0.15222820  0.885558170   0   0
#>   0.545764867   0.13273675  0.607300309   0   0
#>   0.990613569   0.27821927  0.988170573   0   0
#>   0.988170573   0.28028361  0.968202107   0   0
#>  24.121474010   0.60800205  3.700902203   0   0
#>   0.583910978   0.12833457  0.608992254   0   0
#>   0.885558170   0.13606085  0.932691486   0   0
#>   0.771155431   0.28365506  0.794249746   0   0
#>   0.932691486   0.19266047  0.990613569   0   0
#>   0.415781676   0.15856745  0.407572068   0   0
#>   0.596760183   0.07928225  0.673141222   0   0
#>   0.673141222  -0.06891321  0.766264391   0   0
#>   0.766264391  -0.53682851  0.742044373   0   0
#>   0.236682176   0.12601593  0.234565186   0   0
#>   0.339716338   0.05164699  0.352695715   0   0
#>   0.442050170   0.11050667  0.516582980   0   0
#>   0.560264186   0.11050667  0.516582980   0   0
#>   0.408910526   0.15856745  0.407572068   0   0
#>   0.352695715   0.11593530  0.408910526   0   0
#>   0.346826274   0.08824046  0.241910254   0   0
#>   0.234565186   0.10551056  0.227470938   0   0
#>   0.836049194 -20.66902126 31.847255006   0   1
#>   0.828945031  -0.06891321  0.766264391   0   0
#>   0.209057741   0.12601593  0.234565186   0   0
#>   0.742044373  -2.76612344  0.715036927   0   0
#>   0.241910254  -1.60121224  0.587097511   0   0
#>   0.425516832  -0.54362173  0.485691272   0   0
#>   0.485691272   0.08390127  0.560264186   0   0
#>  -0.558494959   0.26476207  0.123283659   0   0
#>   0.227470938  -0.46765999 -0.558494959   0   0
#>   0.715036927  -4.23386288  0.788006018   0   1
#>   0.550832484   0.91694348  0.459377291   0   0
#>   0.486639524   0.87601691  0.450569040   0   0
#>   0.683107476   0.78206092  0.695280796   0   0
#>   0.741398228   0.78586548  0.706471635   0   0
#>   0.729109624   5.30599245  0.884642567   0   1
#>   0.569412895   0.68743736  0.559240914   0   0
#>   0.874110353   0.83864331  0.821970900   0   0
#>   0.730865377   0.86326859  0.874110353   0   0
#>   0.729368401   0.96850798  0.729109624   0   0
#>   0.594961136   0.78401929  0.683107476   0   0
#>   0.427921409   0.77103112  0.523447689   0   0
#>   0.523447689   0.75301740  0.569412895   0   0
#>   0.450569040   0.91694348  0.459377291   0   0
#>   0.551250157   0.77356216  0.423078877   0   0
#>   0.423078877   0.78091910  0.427921409   0   0
#>   0.437994099   0.90296332  0.322616332   0   0
#>   0.881853853   0.86403469  0.889264890   0   0
#>   0.311057408   0.86765665  0.309237117   0   0
#>   0.559240914   0.68083082  0.551250157   0   0
#>   0.671555587   0.63407092  0.681967470   0   0
#>   0.930493915   0.84775596  0.913555325   0   0
#>   1.184113637   6.33304659  4.373346860   0   1
#>   0.313659418   1.15842980  0.377779781   0   0
#>   0.441318372   0.51798374  0.457940379   0   0
#>   0.681967470   0.62795348  0.716517672   0   0
#>   0.207714326   1.90925245  0.388258666   0   0
#>   0.377779781   0.94683499  0.437994099   0   0
#>   0.774917014   0.54166577  0.733332350   0   0
#>   0.313818534   0.73126224  0.274191392   0   0
#>   0.309237117   0.90296332  0.322616332   0   0
#>   0.245003009   0.91433996  0.172461737   0   0
#>   0.303555096   0.76646211  0.313818534   0   0
#>   0.655667758   0.48690787  0.697168144   0   0
#>   0.399517118   0.64995992  0.441318372   0   0
#>   0.554810629   0.54875717  0.634727903   0   0
#>   0.169724795   0.87964635  0.115082823   0   0
#>   0.190254801   0.71927120  0.261442850   0   0
#>   0.261442850   0.70496947  0.255083958   0   0
#>   0.825360383   0.49616549  0.833785649   0   0
#>   0.833785649   0.47923711  0.851415362   0   0
#>   0.879605662   0.67462960  0.851730775   0   0
#>   0.276618673   0.50939676  0.281644049   0   0
#>   0.525060824   0.51798374  0.457940379   0   0
#>   0.594402920   0.44728975  0.568166218   0   0
#>   0.709530947   0.45851021  0.750023885   0   0
#>   0.933094468   0.68027680  0.934180258   0   0
#>   0.272262690   0.52633634  0.276618673   0   0
#>   0.750023885   0.39709777  0.771165170   0   0
#>   0.683109889   0.46799579  0.709530947   0   0
#>   0.684511046   0.41431745  0.683109889   0   0
#>   0.916867406   0.42070746  0.958021004   0   0
#>   0.837416467   0.39709777  0.771165170   0   0
#>   0.900871985   0.47923711  0.851415362   0   0
#>   0.741735265   0.34642397  0.709361870   0   0
#>   0.608990182   0.25400844  0.661012410   0   0
#>   0.848780122   0.37415371  0.892663766   0   0
#>   0.529674852   0.21803368  0.536932438   0   0
#>   0.551421695   0.34115025  0.608990182   0   0
#>   0.709361870   0.36518516  0.684511046   0   0
#>   0.115082823   0.91975564  0.114435128   0   0
#>   0.794249746   0.28264337  0.810812265   0   0
#>   0.958021004   0.39265843  0.943867413   0   0
#>   0.151381823   0.60125413  0.163716642   0   0
#>   0.383608514   0.44715740  0.382636704   0   0
#>   0.379138534   0.50939676  0.281644049   0   0
#>   0.943867413   0.37415371  0.892663766   0   0
#>   1.185265066   0.60800205  3.700902203   0   0
#>   0.127965323   0.41470025  0.181910233   0   0
#>   0.181910233   0.33849438  0.262794055   0   0
#>   0.455119003   0.28795455  0.529674852   0   0
#>   0.382041059   0.29047263  0.387227643   0   0
#>   0.387227643   0.19584866  0.430094264   0   0
#>   0.270485361   0.33849438  0.262794055   0   0
#>   0.302468115   0.29713394  0.270485361   0   0
#>   0.493913092   0.11050667  0.516582980   0   0
#>   0.810812265   0.23572672  0.885635141   0   0
#>   0.698045547   0.14188660  0.771155431   0   0
#>   0.895175309   0.23572672  0.885635141   0   0
#>  -0.155569253   0.54714761  0.047789206   0   0
#>   0.329368674   0.25710358  0.332233012   0   0
#>   0.123283659   0.23959240  0.172568833   0   0
#>   0.332233012   0.23537002  0.340635826   0   0
#>   0.661963255   0.21462904  0.672321105   0   0
#>   0.065478519   0.36844415  0.091978934   0   0
#>   0.432951797   0.17669340  0.493913092   0   0
#>   0.672328752   0.18063612  0.698045547   0   0
#>   0.545764867   0.21803368  0.536932438   0   0
#>   0.990613569  -6.75634000 56.001282150   0   1
#>   0.988170573   0.73193025 24.121474010   0   0
#>  24.121474010   0.74699471 25.717947996   0   1
#>   0.583910978   0.07572578  0.596760183   0   0
#>   0.885558170   0.10380053  0.836049194   0   0
#>   0.771155431   0.09841761  0.828945031   0   0
#>   0.932691486  -3.11015518  7.129589560   0   1
#>   0.415781676   0.23537002  0.340635826   0   0
#>   0.596760183  -0.53682851  0.742044373   0   0
#>   0.339716338   0.12663118  0.236682176   0   0
#>   0.442050170  -0.11832297  0.425516832   0   0
#>   0.560264186   0.08396378  0.583910978   0   0
#>   0.408910526  -0.02245144  0.442050170   0   0
#>   0.172568833   0.18300068  0.209057741   0   0
#>   0.352695715   0.01401878  0.346826274   0   0
#>   0.836049194   0.09841761  0.828945031   0   0
#>   0.241910254   0.10551056  0.227470938   0   0
#>   0.425516832   0.01401878  0.346826274   0   0
#>   0.485691272  -1.60121224  0.587097511   0   0
#>   0.715036927  -1.60121224  0.587097511   0   0
#>   0.486639524   0.83227571  0.583335639   0   0
#>   0.730865377   0.96816090  0.729368401   0   0
#>   0.930493915   1.36308758  1.184113637   0   0
#>   0.207714326  11.41539444  0.128988300   0   1
#>   0.774917014   0.64254418  0.782521742   0   0
#>   0.235479171   0.79509391  0.303555096   0   0
#>   0.655667758   0.44476726  0.594402920   0   0
#>   0.554810629   0.52146897  0.525060824   0   0
#>   0.169724795   0.78902342  0.190254801   0   0
#>   0.933094468   0.47206749  0.900871985   0   0
#>   0.396946783   0.44525270  0.383608514   0   0
#>   0.007551096   0.66292617  0.151381823   0   0
#>   0.370547283   0.28821915  0.302468115   0   0
#>  -0.155569253   0.64555827 -3.192071040   0   1
#>   0.329368674   0.18300068  0.209057741   0   0
#>   0.661963255   0.21628325  0.545764867   0   0
#>   0.065478519   0.30950301 -4.264935242   0   1
#>   0.432951797   0.16945517  0.415781676   0   0
#>   0.672328752   0.07928225  0.673141222   0   0
#>   0.339716338   0.15856745  0.407572068   0   0
#>  -0.558494959  -2.35459977 -2.698476749   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.856204348 0.647714596
#>   [2,] 0.356278740 0.996945982
#>   [3,] 0.317978907 0.406203454
#>   [4,] 0.900287444 0.987738260
#>   [5,] 0.504620615 0.104362036
#>   [6,] 0.156850045 0.617585293
#>   [7,] 0.328638970 0.222795290
#>   [8,] 0.009053839 0.281170657
#>   [9,] 0.766233971 0.241700603
#>  [10,] 0.897230472 0.379940197
#>  [11,] 0.533048291 0.900365551
#>  [12,] 0.863862061 0.673701596
#>  [13,] 0.265788048 0.450797393
#>  [14,] 0.812188327 0.366939986
#>  [15,] 0.083588733 0.366240077
#>  [16,] 0.265942454 0.409920083
#>  [17,] 0.413740851 0.883069814
#>  [18,] 0.200015446 0.980791459
#>  [19,] 0.711154980 0.649134982
#>  [20,] 0.651142951 0.247297607
#>  [21,] 0.552523184 0.701199997
#>  [22,] 0.107216021 0.720477961
#>  [23,] 0.282460565 0.611423391
#>  [24,] 1.000000000 0.614964673
#>  [25,] 0.504771541 0.150841024
#>  [26,] 0.921654213 0.838289577
#>  [27,] 0.823080740 0.000000000
#>  [28,] 0.105434644 0.624797508
#>  [29,] 0.590915138 0.505473225
#>  [30,] 0.056047397 0.931408513
#>  [31,] 0.240140310 0.274802330
#>  [32,] 0.044098795 0.528860143
#>  [33,] 0.337579316 0.910594189
#>  [34,] 0.826888539 0.994089949
#>  [35,] 0.257237259 0.390704196
#>  [36,] 0.825383919 0.245310964
#>  [37,] 0.563793727 0.759036924
#>  [38,] 0.600983660 0.593434858
#>  [39,] 0.148220232 0.628454355
#>  [40,] 0.077582088 0.329650209
#>  [41,] 0.494322153 0.973740172
#>  [42,] 0.998138746 0.578104263
#>  [43,] 0.945335040 0.361648738
#>  [44,] 0.130066638 0.581221536
#>  [45,] 0.182878867 0.325480072
#>  [46,] 0.048750774 0.627436835
#>  [47,] 0.751437407 0.112594683
#>  [48,] 0.393830872 0.583006520
#>  [49,] 0.696692710 0.307109794
#>  [50,] 0.401407006 0.637263008
#>  [51,] 0.000000000 0.236858382
#>  [52,] 0.311902469 0.352628564
#>  [53,] 0.496941906 0.592919999
#>  [54,] 0.801568218 0.844360184
#>  [55,] 0.738052585 0.214087621
#>  [56,] 0.985756339 0.503690316
#>  [57,] 0.826826039 0.231932038
#>  [58,] 0.737271985 0.480755378
#>  [59,] 0.404040790 0.729618022
#>  [60,] 0.516375588 0.755932748
#>  [61,] 0.538575536 0.986141840
#>  [62,] 0.355124791 0.482758260
#>  [63,] 0.311159987 0.659497444
#>  [64,] 0.879486420 0.807559690
#>  [65,] 0.450367574 0.512771153
#>  [66,] 0.203288530 0.771398064
#>  [67,] 0.824808169 0.515060712
#>  [68,] 0.200190942 0.790413146
#>  [69,] 0.366965622 0.800600269
#>  [70,] 0.866296507 0.523671451
#>  [71,] 0.991743643 0.845216993
#>  [72,] 0.086329120 0.453526014
#>  [73,] 0.989337460 0.247099379
#>  [74,] 0.217868887 0.288879394
#>  [75,] 0.768027434 0.243744818
#>  [76,] 0.493759751 0.001326674
#>  [77,] 0.704057454 0.728030573
#>  [78,] 0.146697137 0.129878576
#>  [79,] 0.936659891 0.275889119
#>  [80,] 0.444903918 0.646598461
#>  [81,] 0.814291836 0.786154030
#>  [82,] 0.561208530 0.794255566
#>  [83,] 0.653550937 0.266543761
#>  [84,] 0.343676749 0.303661809
#>  [85,] 0.069331256 0.382545195
#>  [86,] 0.379008711 0.270251878
#>  [87,] 0.336919569 0.876146952
#>  [88,] 0.182671139 0.997739767
#>  [89,] 0.037847357 0.581465316
#>  [90,] 0.929314354 0.929971860
#>  [91,] 0.576851216 0.393807857
#>  [92,] 0.730132237 0.475589803
#>  [93,] 0.033788501 0.916498416
#>  [94,] 0.200563916 1.000000000
#>  [95,] 0.812819018 0.488575293
#>  [96,] 0.428144015 0.799634479
#>  [97,] 0.285352517 0.708826039
#>  [98,] 0.530034166 0.919813450
#>  [99,] 0.615157588 0.067376796
#> [100,] 0.262894965 0.005049415
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,24,12), [9,7,2]
#> 2: (42,24,1), [1,3,45]
#> 3: (70,42,1), [2,6,4]
#> 4: (56,42,70), [3,11,46]
#> 5: (70,67,95), [22,23,6]
#> 6: (67,70,1), [3,16,5]
#> 7: (19,1,12), [1,10,16]
#> 8: (12,64,81), [18,20,15]
#> 9: (71,12,24), [1,0,15]
#> 10: (19,12,77), [20,41,7]
#> 11: (70,10,56), [26,4,23]
#> 12: (58,67,19), [16,31,22]
#> 13: (64,26,54), [19,18,14]
#> 14: (64,71,26), [27,13,15]
#> 15: (71,64,12), [8,9,14]
#> 16: (19,67,1), [6,7,12]
#> 17: (81,54,77), [33,20,18]
#> 18: (64,54,81), [17,8,13]
#> 19: (26,90,54), [34,13,27]
#> 20: (12,81,77), [17,10,8]
#> 21: (58,14,95), [29,22,25]
#> 22: (58,95,67), [5,12,21]
#> 23: (70,95,10), [29,11,5]
#> 24: (58,38,92), [30,25,31]
#> 25: (58,92,14), [40,21,24]
#> 26: (43,56,10), [11,38,47]
#> 27: (26,71,90), [36,19,14]
#> 28: (10,14,36), [55,51,29]
#> 29: (95,14,10), [28,23,21]
#> 30: (92,38,29), [63,39,24]
#> 31: (58,19,38), [32,24,12]
#> 32: (21,38,19), [31,41,42]
#> 33: (54,82,77), [49,17,70]
#> 34: (90,34,54), [43,19,35]
#> 35: (90,4,34), [0,34,36]
#> 36: (4,90,71), [27,0,35]
#> 37: (43,79,73), [52,47,38]
#> 38: (10,79,43), [37,26,51]
#> 39: (91,92,29), [30,60,59]
#> 40: (92,49,14), [50,25,59]
#> 41: (21,19,77), [10,44,32]
#> 42: (53,38,21), [32,56,63]
#> 43: (54,34,61), [117,80,34]
#> 44: (77,37,21), [57,41,49]
#> 45: (73,24,42), [2,46,0]
#> 46: (73,42,56), [4,47,45]
#> 47: (43,73,56), [46,26,37]
#> 48: (37,82,60), [68,57,49]
#> 49: (82,37,77), [44,33,48]
#> 50: (75,14,49), [40,62,55]
#> 51: (10,36,79), [53,38,28]
#> 52: (79,57,73), [89,37,53]
#> 53: (79,36,57), [54,52,51]
#> 54: (36,75,57), [61,53,55]
#> 55: (75,36,14), [28,50,54]
#> 56: (80,53,21), [42,75,74]
#> 57: (37,60,21), [75,44,48]
#> 58: (83,49,91), [59,72,64]
#> 59: (91,49,92), [40,39,58]
#> 60: (29,65,91), [119,39,73]
#> 61: (9,57,75), [54,62,66]
#> 62: (49,9,75), [61,50,67]
#> 63: (53,29,38), [30,42,73]
#> 64: (55,49,83), [58,71,67]
#> 65: (57,55,47), [90,107,66]
#> 66: (55,57,9), [61,67,65]
#> 67: (55,9,49), [62,64,66]
#> 68: (96,60,82), [48,69,81]
#> 69: (96,82,11), [70,82,68]
#> 70: (54,11,82), [69,33,77]
#> 71: (20,55,83), [64,78,90]
#> 72: (25,83,91), [58,91,78]
#> 73: (29,53,65), [83,60,63]
#> 74: (53,80,48), [86,83,56]
#> 75: (80,21,60), [57,76,56]
#> 76: (59,80,60), [75,81,84]
#> 77: (98,11,54), [70,80,94]
#> 78: (25,20,83), [71,72,102]
#> 79: (41,98,61), [80,116,88]
#> 80: (98,54,61), [43,79,77]
#> 81: (59,60,96), [68,92,76]
#> 82: (11,17,96), [93,69,94]
#> 83: (53,48,65), [85,73,74]
#> 84: (59,50,80), [86,76,87]
#> 85: (62,65,48), [83,104,96]
#> 86: (80,50,48), [98,74,84]
#> 87: (59,63,50), [98,84,106]
#> 88: (17,98,41), [79,110,94]
#> 89: (57,27,73), [0,52,107]
#> 90: (47,55,20), [71,111,65]
#> 91: (91,86,25), [123,72,113]
#> 92: (96,69,59), [95,81,93]
#> 93: (96,17,69), [100,92,82]
#> 94: (11,98,17), [88,82,77]
#> 95: (59,69,97), [99,106,92]
#> 96: (3,65,62), [85,124,112]
#> 97: (23,48,63), [98,105,104]
#> 98: (63,48,50), [86,87,97]
#> 99: (69,66,97), [125,95,109]
#> 100: (87,69,17), [93,101,134]
#> 101: (17,33,87), [136,100,115]
#> 102: (99,20,25), [78,122,111]
#> 103: (13,62,23), [104,141,124]
#> 104: (23,62,48), [85,97,103]
#> 105: (63,97,23), [145,97,106]
#> 106: (59,97,63), [105,87,95]
#> 107: (57,47,27), [108,89,65]
#> 108: (47,99,27), [137,107,111]
#> 109: (68,66,69), [99,134,158]
#> 110: (17,41,2), [116,115,88]
#> 111: (99,47,20), [90,102,108]
#> 112: (65,3,84), [120,119,96]
#> 113: (91,84,86), [132,91,119]
#> 114: (18,33,2), [115,154,136]
#> 115: (17,2,33), [114,101,110]
#> 116: (2,41,61), [79,117,110]
#> 117: (2,61,34), [43,155,116]
#> 118: (7,5,25), [122,123,139]
#> 119: (65,84,91), [113,60,112]
#> 120: (52,84,3), [112,126,129]
#> 121: (16,3,13), [124,130,127]
#> 122: (5,99,25), [102,118,138]
#> 123: (7,25,86), [91,132,118]
#> 124: (13,3,62), [96,103,121]
#> 125: (97,66,39), [135,133,99]
#> 126: (35,52,3), [120,127,143]
#> 127: (35,3,16), [121,130,126]
#> 128: (84,31,7), [171,132,129]
#> 129: (84,52,31), [140,128,120]
#> 130: (35,16,13), [121,148,127]
#> 131: (72,13,44), [141,168,148]
#> 132: (7,86,84), [113,128,123]
#> 133: (6,97,39), [125,144,145]
#> 134: (68,69,87), [100,151,109]
#> 135: (22,39,66), [125,158,150]
#> 136: (87,33,18), [114,151,101]
#> 137: (76,27,99), [108,138,0]
#> 138: (76,99,5), [122,146,137]
#> 139: (100,5,7), [118,142,146]
#> 140: (74,31,52), [129,143,177]
#> 141: (23,44,13), [131,103,152]
#> 142: (78,100,7), [139,171,183]
#> 143: (74,52,35), [126,147,140]
#> 144: (44,6,39), [133,149,152]
#> 145: (6,23,97), [105,133,152]
#> 146: (100,76,5), [138,139,0]
#> 147: (35,45,74), [164,143,160]
#> 148: (72,35,13), [130,131,160]
#> 149: (28,44,39), [144,150,156]
#> 150: (22,28,39), [149,135,162]
#> 151: (87,18,68), [157,134,136]
#> 152: (23,6,44), [144,141,145]
#> 153: (88,18,94), [154,0,159]
#> 154: (94,18,2), [114,155,153]
#> 155: (94,2,34), [117,0,154]
#> 156: (89,44,28), [149,161,169]
#> 157: (68,18,30), [159,175,151]
#> 158: (22,66,68), [109,176,135]
#> 159: (30,18,88), [153,0,157]
#> 160: (35,72,45), [166,147,148]
#> 161: (89,28,46), [162,178,156]
#> 162: (28,22,46), [163,161,150]
#> 163: (22,93,46), [178,162,176]
#> 164: (45,40,74), [174,147,165]
#> 165: (45,15,40), [172,164,166]
#> 166: (15,45,72), [160,170,165]
#> 167: (85,72,32), [168,180,170]
#> 168: (44,32,72), [167,131,169]
#> 169: (89,32,44), [168,156,181]
#> 170: (85,15,72), [166,167,172]
#> 171: (78,7,31), [128,177,142]
#> 172: (40,15,85), [170,173,165]
#> 173: (85,8,40), [179,172,180]
#> 174: (74,40,78), [184,177,164]
#> 175: (68,30,93), [0,176,157]
#> 176: (68,93,22), [163,158,175]
#> 177: (31,74,78), [174,171,140]
#> 178: (46,93,89), [185,161,163]
#> 179: (40,8,51), [182,184,173]
#> 180: (85,32,8), [181,173,167]
#> 181: (8,32,89), [169,182,180]
#> 182: (89,51,8), [179,181,185]
#> 183: (51,100,78), [142,184,0]
#> 184: (40,51,78), [183,174,179]
#> 185: (89,93,51), [0,182,178]
#> boundary nodes:  24 71 4 34 94 88 30 93 51 100 76 27 73 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.148268556 0.147477954 0.158085330 0.142636976 0.029072618 0.124453018 
#>                                                                         
#> 0.027091783 0.068618773 0.168787293 0.152154268 0.175763925 0.120239338 
#>                                                                         
#> 0.070430940 0.134766872 0.136318646 0.151792383 0.059580591 0.153780474 
#>                                                                         
#> 0.124619241 0.121636942 0.137574990 0.174833670 0.136154267 0.105265396 
#>                                                                         
#> 0.122342726 0.086030064 0.088536003 0.123451300 0.157453373 0.151855594 
#>                                                                         
#> 0.073673219 0.060031454 0.086197276 0.142388271 0.130072631 0.079214198 
#>                                                                         
#> 0.118159819 0.288422538 0.058924828 0.036907372 0.075437123 0.256615926 
#>                                                                         
#> 0.058979921 0.143649946 0.115400891 0.163217677 0.013456425 0.059973475 
#>                                                                         
#> 0.121712105 0.065592093 0.147913953 0.171766455 0.173638641 0.002719435 
#>                                                                         
#> 0.059218372 0.102371724 0.061374454 0.095469102 0.109783096 0.099459415 
#>                                                                         
#> 0.148590031 0.092698468 0.081562192 0.130622302 0.274298291 0.019396207 
#>                                                                         
#> 0.066876081 0.298776152 0.098461423 0.084669394 0.090163215 0.044487437 
#>                                                                         
#> 0.054782884 0.092945654 0.064679511 0.297824445 0.093038636 0.173422148 
#>                                                                         
#> 0.080081505 0.094811117 0.121959952 0.122813714 0.099859709 0.112629200 
#>                                                                         
#> 0.103197656 0.034453554 0.175295357 0.147766051 0.107461036 0.097445571 
#>                                                                         
#> 0.055671679 0.133455348 0.218567312 0.166261719 0.139980333 0.167940036 
#>                                                                         
#> 0.105712667 0.048626855 0.088353272 0.045958277 0.046479233 0.249993285 
#>                                                                         
#> 0.068647689 0.138388427 0.085090864 0.153184273 0.053918381 0.052169011 
#>                                                                         
#> 0.102648649 0.105861845 0.040877602 0.188230598 0.158949272 0.081302316 
#>                                                                         
#> 0.154439227 0.116560414 0.212122547 0.227454479 0.066618608 0.013878407 
#>                                                                         
#> 0.103606198 0.050635078 0.060698504 0.050601306 0.042941577 0.190378394 
#>                                                                         
#> 0.045162747 0.019216370 0.157096103 0.470618466 0.050055973 0.152202085 
#>                                                                         
#> 0.019265732 0.024250158 0.160367131 0.056745283 0.109885560 0.289448558 
#>                                                                         
#> 0.146091222 0.037079620 0.086363369 0.100658757 0.087328945 0.021659468 
#>                                                                         
#> 0.083942765 0.211389676 0.026791220 0.209321809 0.174202979 0.335057686 
#>                                                                         
#> 0.045227754 0.250156412 0.052975321 0.170541065 0.181561915 0.680479418 
#>                                                                         
#> 0.121119351 0.063965476 0.154670525 0.122893275 0.230400301 0.170393043 
#>                                                                         
#> 0.086171471 0.052177090 0.213942120 0.145056323 0.075950677 0.094018294 
#>                                                                         
#> 0.147022391 0.008812443 0.136258584 0.051464812 0.092001732 0.152600647 
#>                                                                         
#> 0.176836571 0.166957516 0.120839467 0.169341512 0.122710100 0.112547539 
#>                                                                         
#> 0.153891243 0.331121875 0.147681047 0.047519636 0.095413642 0.111271316 
#>                                                                         
#> 0.118303424 0.130871058 0.120678226 0.173733583 0.104043028 0.141155411 
#>                                                                         
#> 0.039454623 0.101802574 0.133173188 0.145436090 0.245526480 0.253433150 
#>                                                                         
#> 0.140736909 0.103586497 0.115375785 0.281822528 0.188473956 0.074049103 
#>                                                                         
#> 0.092531307 0.092392561 0.121303173 0.061186015 0.120554584 0.120495696 
#>                                                                         
#> 0.055989029 0.077132583 0.161488959 0.114938522 0.116377571 0.231962276 
#>                                                                         
#> 0.167085521 0.127552639 0.183484181 0.234754446 0.233254661 0.080982533 
#>                                                                         
#> 0.182616678 0.190263236 0.069204325 0.094881772 0.062687916 0.021095752 
#>                                                                         
#> 0.082252834 0.107483318 0.134978209 0.161384642 0.108732626 0.172317184 
#>                                                                         
#> 0.138202943 0.113605772 0.204294753 0.109170321 0.157599997 0.261331756 
#>                                                                         
#> 0.179479671 0.100745668 0.155357848 0.018034970 0.155744770 0.626352507 
#>                                                                         
#> 0.080285243 0.201635947 0.116341163 0.142945469 0.098910852 0.047246852 
#>                                                                         
#> 0.105379323 0.148474716 0.092219603 0.072987697 0.172437006 0.053534623 
#>                                                                         
#> 0.208775640 0.120951658 0.117941358 0.301671922 0.346679047 0.042372480 
#>                                                                         
#> 0.118405053 0.064649195 0.368019788 0.035313396 0.057377863 0.074761818 
#>                                                                         
#> 0.128366298 0.090549154 0.019680085 0.169951269 0.143585676 0.058372493 
#>                                                                         
#> 0.329323661 0.026347296 0.125761567 0.230894799 0.182088366 0.095697034 
#>                         
#> 0.107330838 0.350498445 
#> 
#> $alpha
#> [1] 0.2814009
#> 
#> attr(,"class")
#> [1] "scree"