Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1          y1         x2          y2           mx1
#>    79   97 0.17917927 0.798069152 0.15299606 0.825541117  0.1755025594
#>     2   50 0.15784677 0.721725445 0.22105670 0.721083940  0.1897820744
#>    79    1 0.17917927 0.798069152 0.17036079 0.798355414  0.1734812028
#>     2    1 0.15784677 0.721725445 0.17036079 0.798355414  0.1202415620
#>    44   85 0.31195484 0.759108825 0.21504412 0.910509614  0.2617317321
#>    83   49 0.06903141 0.872834801 0.06129480 0.764434083  0.0926973739
#>    97   85 0.15299606 0.825541117 0.21504412 0.910509614  0.1398088730
#>     1   97 0.17036079 0.798355414 0.15299606 0.825541117  0.1154991339
#>     2   49 0.15784677 0.721725445 0.06129480 0.764434083  0.1066250558
#>    50   44 0.22105670 0.721083940 0.31195484 0.759108825  0.2475608968
#>    35   58 0.28951936 0.621634572 0.29065366 0.707249242  0.2460866386
#>    85   59 0.21504412 0.910509614 0.28878545 0.947824670  0.2326683627
#>    42   49 0.12262931 0.685592500 0.06129480 0.764434083  0.0813028594
#>    83   62 0.06903141 0.872834801 0.03416569 0.938172356 -0.0137877913
#>    93   72 0.00107170 0.952117394 0.01399636 0.718147388 -0.0594983849
#>    42   36 0.12262931 0.685592500 0.02960219 0.714746005  0.0486164667
#>    36   49 0.02960219 0.714746005 0.06129480 0.764434083  0.0291142735
#>    62   85 0.03416569 0.938172356 0.21504412 0.910509614  0.1416038710
#>    72   49 0.01399636 0.718147388 0.06129480 0.764434083 -0.0457933092
#>    62   41 0.03416569 0.938172356 0.28215426 1.000000000  0.1096667453
#>     9   15 0.38749032 0.752828567 0.40285363 0.753863191  0.3928447179
#>    44    9 0.31195484 0.759108825 0.38749032 0.752828567  0.3537780368
#>    44   58 0.31195484 0.759108825 0.29065366 0.707249242  0.3462914267
#>     9   58 0.38749032 0.752828567 0.29065366 0.707249242  0.3704484955
#>    42   98 0.12262931 0.685592500 0.26787097 0.606248929  0.1936217468
#>    98   35 0.26787097 0.606248929 0.28951936 0.621634572  0.2436045029
#>    36   72 0.02960219 0.714746005 0.01399636 0.718147388 -0.0007393129
#>    33   44 0.41498122 0.820792646 0.31195484 0.759108825  0.3437897880
#>     9   35 0.38749032 0.752828567 0.28951936 0.621634572  0.3912185046
#>    59   44 0.28878545 0.947824670 0.31195484 0.759108825  0.3237231923
#>    59   33 0.28878545 0.947824670 0.41498122 0.820792646  0.3669908750
#>    60   98 0.14134759 0.551806441 0.26787097 0.606248929  0.1843842580
#>    41   59 0.28215426 1.000000000 0.28878545 0.947824670  0.3951753921
#>    21   33 0.45884640 0.893689086 0.41498122 0.820792646  0.5085858748
#>    15   71 0.40285363 0.753863191 0.49436214 0.667672004  0.4579151357
#>    22   46 0.27666142 0.541416004 0.27647413 0.520163133  0.3652172197
#>    22   60 0.27666142 0.541416004 0.14134759 0.551806441  0.2078361047
#>    61   34 0.02704967 0.510602696 0.05287777 0.493927366  0.0750437695
#>    60   42 0.14134759 0.551806441 0.12262931 0.685592500  0.0564663606
#>    23   46 0.18456754 0.462548785 0.27647413 0.520163133  0.2071792918
#>    73   15 0.41467106 0.815877706 0.40285363 0.753863191  0.4900175402
#>    71   35 0.49436214 0.667672004 0.28951936 0.621634572  0.4029320499
#>    34   23 0.05287777 0.493927366 0.18456754 0.462548785  0.1207482783
#>    31   74 0.51017101 0.683175399 0.57994556 0.753167793  0.4921465294
#>    41   21 0.28215426 1.000000000 0.45884640 0.893689086  0.3983013154
#>    39   12 0.54035541 0.975769306 0.51449948 0.999085547  0.4892102866
#>    33   73 0.41498122 0.820792646 0.41467106 0.815877706  0.5079089234
#>    19   39 0.60596813 0.775706170 0.54035541 0.975769306  0.5625202275
#>    74   19 0.57994556 0.753167793 0.60596813 0.775706170  0.5322612210
#>    35   11 0.28951936 0.621634572 0.43126797 0.469928725  0.3913818384
#>    38   88 0.48556525 0.513867738 0.53416589 0.492592609  0.5467642853
#>    22   11 0.27666142 0.541416004 0.43126797 0.469928725  0.3817273843
#>    46   25 0.27647413 0.520163133 0.23130586 0.366725564  0.2810570421
#>    67   31 0.62846196 0.641495717 0.51017101 0.683175399  0.5549863217
#>    23   25 0.18456754 0.462548785 0.23130586 0.366725564  0.2620640598
#>    11   20 0.43126797 0.469928725 0.31976105 0.359949877  0.3839325402
#>    11   46 0.43126797 0.469928725 0.27647413 0.520163133  0.3396688931
#>    88   67 0.53416589 0.492592609 0.62846196 0.641495717  0.5483937181
#>    11   68 0.43126797 0.469928725 0.46313464 0.408142881  0.4852952692
#>    87   27 0.00000000 0.362680383 0.22540391 0.132741786  0.1176147135
#>    61   87 0.02704967 0.510602696 0.00000000 0.362680383 -0.0007225237
#>    61   72 0.02704967 0.510602696 0.01399636 0.718147388 -0.7078896585
#>    30   20 0.27094723 0.332618012 0.31976105 0.359949877  0.2767695970
#>    68   88 0.46313464 0.408142881 0.53416589 0.492592609  0.4899341008
#>    20    6 0.31976105 0.359949877 0.32334952 0.338241401  0.3764448996
#>    30    6 0.27094723 0.332618012 0.32334952 0.338241401  0.2961328304
#>    84   92 0.73573316 0.893729048 0.67302977 0.911249084  0.6849057929
#>    67   63 0.62846196 0.641495717 0.68812767 0.615751138  0.6986959634
#>    68    5 0.46313464 0.408142881 0.41827390 0.319873297  0.4573370168
#>    68   20 0.46313464 0.408142881 0.31976105 0.359949877  0.3894229142
#>    10    5 0.37602833 0.290050978 0.41827390 0.319873297  0.3728189413
#>    84  100 0.73573316 0.893729048 0.87166482 0.972873474  0.7424051955
#>    92   19 0.67302977 0.911249084 0.60596813 0.775706170  0.6791216714
#>    30   94 0.27094723 0.332618012 0.34999605 0.268097367  0.3020228340
#>    94   10 0.34999605 0.268097367 0.37602833 0.290050978  0.3412316533
#>    64   52 0.66328201 0.513660998 0.68769353 0.551736014  0.6175291749
#>    63   52 0.68812767 0.615751138 0.68769353 0.551736014  0.7048820681
#>    27   30 0.22540391 0.132741786 0.27094723 0.332618012  0.1664823032
#>    19   63 0.60596813 0.775706170 0.68812767 0.615751138  0.7056490056
#>    18   96 0.50461988 0.331818110 0.58633945 0.305826610  0.5721000534
#>    63   43 0.68812767 0.615751138 0.73485449 0.563275863  0.7262068096
#>    96   64 0.58633945 0.305826610 0.66328201 0.513660998  0.6132641073
#>    10   32 0.37602833 0.290050978 0.34266171 0.224802441  0.4432533374
#>    32   94 0.34266171 0.224802441 0.34999605 0.268097367  0.2761720067
#>    84   40 0.73573316 0.893729048 0.76140231 0.853897987  0.7918067663
#>    45   53 0.78505026 0.647601454 0.79989055 0.845624476  0.7590124695
#>    63   80 0.68812767 0.615751138 0.76194467 0.584034209  0.7375006261
#>    91   89 0.64375476 0.296947795 0.74419213 0.428727221  0.6362080086
#>    40   19 0.76140231 0.853897987 0.60596813 0.775706170  0.7190321629
#>    18    5 0.50461988 0.331818110 0.41827390 0.319873297  0.4742989791
#>    66   75 0.30211779 0.170344573 0.33756237 0.183582125  0.3081962715
#>    43   64 0.73485449 0.563275863 0.66328201 0.513660998  0.7290124482
#>    30   66 0.27094723 0.332618012 0.30211779 0.170344573  0.2574780402
#>    32   75 0.34266171 0.224802441 0.33756237 0.183582125  0.4311371736
#>    27   55 0.22540391 0.132741786 0.29818312 0.153175115  0.2535142886
#>    82   95 0.83162585 0.637773550 0.85135158 0.658832758  0.8147983270
#>    45   80 0.78505026 0.647601454 0.76194467 0.584034209  0.8002572862
#>    43   89 0.73485449 0.563275863 0.74419213 0.428727221  0.7455408880
#>    80   43 0.76194467 0.584034209 0.73485449 0.563275863  0.7724947568
#>    95   53 0.85135158 0.658832758 0.79989055 0.845624476  0.8025072746
#>    80   65 0.76194467 0.584034209 0.80918163 0.519524860  0.7973825457
#>    82   45 0.83162585 0.637773550 0.78505026 0.647601454  0.8006821838
#>    96   91 0.58633945 0.305826610 0.64375476 0.296947795  0.6314249127
#>    77   96 0.61617783 0.205913084 0.58633945 0.305826610  0.5205343427
#>    82   48 0.83162585 0.637773550 0.83550071 0.579402820  0.8434761499
#>    39   92 0.54035541 0.975769306 0.67302977 0.911249084  0.7075536168
#>    84   53 0.73573316 0.893729048 0.79989055 0.845624476  0.8088646559
#>   100   53 0.87166482 0.972873474 0.79989055 0.845624476  0.8817545142
#>    91   81 0.64375476 0.296947795 0.65736202 0.290457923  0.6782823338
#>     3   90 0.85713494 0.581539292 0.91203957 0.641903157  0.8736817170
#>     3   48 0.85713494 0.581539292 0.83550071 0.579402820  0.8490624062
#>    48   65 0.83550071 0.579402820 0.80918163 0.519524860  0.8178508881
#>    57   89 0.77017137 0.363658302 0.74419213 0.428727221  0.7201002567
#>    57   78 0.77017137 0.363658302 0.83766611 0.524962424  0.8412857570
#>    65   89 0.80918163 0.519524860 0.74419213 0.428727221  0.8395967451
#>    81   77 0.65736202 0.290457923 0.61617783 0.205913084  0.6912905756
#>    55   29 0.29818312 0.153175115 0.32230567 0.062549420  0.3553336745
#>     8   13 0.97997541 0.760056758 0.96536426 0.849674222  0.8840281777
#>     3   78 0.85713494 0.581539292 0.83766611 0.524962424  0.9283761918
#>    27   29 0.22540391 0.132741786 0.32230567 0.062549420  0.2743309120
#>    51   56 0.74006973 0.332167096 0.72867731 0.294424558  0.8107184812
#>    29   99 0.32230567 0.062549420 0.50442667 0.134005351  0.4074753822
#>    90    8 0.91203957 0.641903157 0.97997541 0.760056758  0.9027523518
#>     7   70 0.17671769 0.026384631 0.27590595 0.014597437  0.2316668254
#>    37    7 0.08902115 0.003266755 0.17671769 0.026384631  0.1041038873
#>    95    8 0.85135158 0.658832758 0.97997541 0.760056758  0.8719491162
#>    16    8 0.98990998 0.592035104 0.97997541 0.760056758  0.9889595323
#>    29   70 0.32230567 0.062549420 0.27590595 0.014597437  0.3701532561
#>    87   47 0.00000000 0.362680383 0.02831269 0.000000000  0.0476642345
#>   100   13 0.87166482 0.972873474 0.96536426 0.849674222  0.9240664282
#>    47   37 0.02831269 0.000000000 0.08902115 0.003266755  0.0489595357
#>    24    8 1.00000000 0.931993500 0.97997541 0.760056758  1.9563262988
#>    16   78 0.98990998 0.592035104 0.83766611 0.524962424  0.9588264974
#>    56   81 0.72867731 0.294424558 0.65736202 0.290457923  0.6968440473
#>    77   76 0.61617783 0.205913084 0.62356571 0.095798510  0.6350435327
#>    77   99 0.61617783 0.205913084 0.50442667 0.134005351  0.5745511593
#>    29   28 0.32230567 0.062549420 0.46552480 0.012149823  0.4144151845
#>    14   69 0.78076938 0.213758161 0.89742362 0.331255674  0.8196331703
#>    28   76 0.46552480 0.012149823 0.62356571 0.095798510  0.5444904162
#>    70   37 0.27590595 0.014597437 0.08902115 0.003266755  0.3376240874
#>    17   14 0.71037641 0.147299428 0.78076938 0.213758161  0.7058927427
#>    86   26 0.67708195 0.113028631 0.69086843 0.036936092  0.7400680585
#>    17   77 0.71037641 0.147299428 0.61617783 0.205913084  0.6567396280
#>    57   51 0.77017137 0.363658302 0.74006973 0.332167096  0.8141938913
#>    17   26 0.71037641 0.147299428 0.69086843 0.036936092  0.7796369417
#>    17    4 0.71037641 0.147299428 0.85105617 0.011234569  0.8113108668
#>    28   26 0.46552480 0.012149823 0.69086843 0.036936092  0.5829379472
#>    16   69 0.98990998 0.592035104 0.89742362 0.331255674  1.3629302154
#>    54   69 0.94257511 0.112121343 0.89742362 0.331255674  0.8953896312
#>     4   54 0.85105617 0.011234569 0.94257511 0.112121343  0.8339100907
#>    28   47 0.46552480 0.012149823 0.02831269 0.000000000  0.6259011392
#>    85   79 0.21504412 0.910509614 0.17917927 0.798069152  0.2072650261
#>    79   50 0.17917927 0.798069152 0.22105670 0.721083940  0.1897820744
#>    49    1 0.06129480 0.764434083 0.17036079 0.798355414  0.1202415620
#>    97   49 0.15299606 0.825541117 0.06129480 0.764434083  0.0926973739
#>    83   85 0.06903141 0.872834801 0.21504412 0.910509614  0.1398088730
#>    42    2 0.12262931 0.685592500 0.15784677 0.721725445  0.1887913654
#>    79   44 0.17917927 0.798069152 0.31195484 0.759108825  0.2475608968
#>    58   50 0.29065366 0.707249242 0.22105670 0.721083940  0.2627800227
#>    93   62 0.00107170 0.952117394 0.03416569 0.938172356 -0.0137877913
#>    83   72 0.06903141 0.872834801 0.01399636 0.718147388 -0.0594983849
#>    61   36 0.02704967 0.510602696 0.02960219 0.714746005  0.0486164667
#>    41   85 0.28215426 1.000000000 0.21504412 0.910509614  0.1416038710
#>    93   41 0.00107170 0.952117394 0.28215426 1.000000000  0.1096667453
#>    73    9 0.41467106 0.815877706 0.38749032 0.752828567  0.3537780368
#>    50   98 0.22105670 0.721083940 0.26787097 0.606248929  0.1936217468
#>    50   35 0.22105670 0.721083940 0.28951936 0.621634572  0.2436045029
#>    73   44 0.41467106 0.815877706 0.31195484 0.759108825  0.3437897880
#>    21   59 0.45884640 0.893689086 0.28878545 0.947824670  0.3951753921
#>    74   33 0.57994556 0.753167793 0.41498122 0.820792646  0.5085858748
#>    31   71 0.51017101 0.683175399 0.49436214 0.667672004  0.4579151357
#>    71    9 0.49436214 0.667672004 0.38749032 0.752828567  0.4014136678
#>    22   98 0.27666142 0.541416004 0.26787097 0.606248929  0.3041305600
#>    46   60 0.27647413 0.520163133 0.14134759 0.551806441  0.2078361047
#>    60   34 0.14134759 0.551806441 0.05287777 0.493927366  0.0750437695
#>    61   42 0.02704967 0.510602696 0.12262931 0.685592500  0.0564663606
#>    31   15 0.51017101 0.683175399 0.40285363 0.753863191  0.4900175402
#>    38   35 0.48556525 0.513867738 0.28951936 0.621634572  0.4029320499
#>    60   23 0.14134759 0.551806441 0.18456754 0.462548785  0.1207482783
#>    73   74 0.41467106 0.815877706 0.57994556 0.753167793  0.4921465294
#>    12   21 0.51449948 0.999085547 0.45884640 0.893689086  0.3983013154
#>    21   39 0.45884640 0.893689086 0.54035541 0.975769306  0.5625202275
#>    21   19 0.45884640 0.893689086 0.60596813 0.775706170  0.5322612210
#>    38   11 0.48556525 0.513867738 0.43126797 0.469928725  0.3913818384
#>    71   88 0.49436214 0.667672004 0.53416589 0.492592609  0.5467642853
#>    20   25 0.31976105 0.359949877 0.23130586 0.366725564  0.2810570421
#>    87   23 0.00000000 0.362680383 0.18456754 0.462548785  0.1155675741
#>    87   34 0.00000000 0.362680383 0.05287777 0.493927366  0.0997835210
#>    20   46 0.31976105 0.359949877 0.27647413 0.520163133  0.3396688931
#>    71   67 0.49436214 0.667672004 0.62846196 0.641495717  0.5483937181
#>    38   68 0.48556525 0.513867738 0.46313464 0.408142881  0.4852952692
#>    67   74 0.62846196 0.641495717 0.57994556 0.753167793  0.6382617653
#>    25   27 0.23130586 0.366725564 0.22540391 0.132741786  0.1176147135
#>    87   72 0.00000000 0.362680383 0.01399636 0.718147388 -0.7078896585
#>    87   93 0.00000000 0.362680383 0.00107170 0.952117394 -3.1076046778
#>     5    6 0.41827390 0.319873297 0.32334952 0.338241401  0.3764448996
#>   100   39 0.87166482 0.972873474 0.54035541 0.975769306  0.7078736667
#>    40   92 0.76140231 0.853897987 0.67302977 0.911249084  0.6849057929
#>    64   88 0.66328201 0.513660998 0.53416589 0.492592609  0.5890987319
#>    18   68 0.50461988 0.331818110 0.46313464 0.408142881  0.5509074912
#>     5   20 0.41827390 0.319873297 0.31976105 0.359949877  0.3894229142
#>    92  100 0.67302977 0.911249084 0.87166482 0.972873474  0.7424051955
#>     6   94 0.32334952 0.338241401 0.34999605 0.268097367  0.3020228340
#>     6   10 0.32334952 0.338241401 0.37602833 0.290050978  0.3412316533
#>    67   52 0.62846196 0.641495717 0.68769353 0.551736014  0.6175291749
#>    43   52 0.73485449 0.563275863 0.68769353 0.551736014  0.7048820681
#>    25   30 0.23130586 0.366725564 0.27094723 0.332618012  0.1664823032
#>    45   63 0.78505026 0.647601454 0.68812767 0.615751138  0.7056490056
#>    88   96 0.53416589 0.492592609 0.58633945 0.305826610  0.5721000534
#>    99   10 0.50442667 0.134005351 0.37602833 0.290050978  0.4544781894
#>    99   32 0.50442667 0.134005351 0.34266171 0.224802441  0.4432533374
#>    53   40 0.79989055 0.845624476 0.76140231 0.853897987  0.7918067663
#>    64   89 0.66328201 0.513660998 0.74419213 0.428727221  0.6362080086
#>    45   19 0.78505026 0.647601454 0.60596813 0.775706170  0.7190321629
#>    99    5 0.50442667 0.134005351 0.41827390 0.319873297  0.4742989791
#>    99   18 0.50442667 0.134005351 0.50461988 0.331818110  0.5181509323
#>    32   66 0.34266171 0.224802441 0.30211779 0.170344573  0.2574780402
#>    99   75 0.50442667 0.134005351 0.33756237 0.183582125  0.4311371736
#>    55   66 0.29818312 0.153175115 0.30211779 0.170344573  0.3278925341
#>    45   95 0.78505026 0.647601454 0.85135158 0.658832758  0.8147983270
#>    48   80 0.83550071 0.579402820 0.76194467 0.584034209  0.8002572862
#>    65   43 0.80918163 0.519524860 0.73485449 0.563275863  0.7724947568
#>    48   45 0.83550071 0.579402820 0.78505026 0.647601454  0.8006821838
#>    64   91 0.66328201 0.513660998 0.64375476 0.296947795  0.6314249127
#>    99   96 0.50442667 0.134005351 0.58633945 0.305826610  0.5205343427
#>    13   53 0.96536426 0.849674222 0.79989055 0.845624476  0.8817545142
#>    51   81 0.74006973 0.332167096 0.65736202 0.290457923  0.6782823338
#>    51   91 0.74006973 0.332167096 0.64375476 0.296947795  0.6666399746
#>    91   77 0.64375476 0.296947795 0.61617783 0.205913084  0.6083350139
#>    95   90 0.85135158 0.658832758 0.91203957 0.641903157  0.8736817170
#>     3   82 0.85713494 0.581539292 0.83162585 0.637773550  0.8702295852
#>    78   48 0.83766611 0.524962424 0.83550071 0.579402820  0.8490624062
#>    78   65 0.83766611 0.524962424 0.80918163 0.519524860  0.8178508881
#>    51   89 0.74006973 0.332167096 0.74419213 0.428727221  0.7201002567
#>    75   29 0.33756237 0.183582125 0.32230567 0.062549420  0.3553336745
#>    69   57 0.89742362 0.331255674 0.77017137 0.363658302  0.8532079705
#>    14   56 0.78076938 0.213758161 0.72867731 0.294424558  0.8107184812
#>    27   70 0.22540391 0.132741786 0.27590595 0.014597437  0.2316668254
#>    27    7 0.22540391 0.132741786 0.17671769 0.026384631  0.1041038873
#>    53    8 0.79989055 0.845624476 0.97997541 0.760056758  0.8719491162
#>    28   70 0.46552480 0.012149823 0.27590595 0.014597437  0.3701532561
#>    27   47 0.22540391 0.132741786 0.02831269 0.000000000  0.0476642345
#>    24   13 1.00000000 0.931993500 0.96536426 0.849674222  0.9240664282
#>    27   37 0.22540391 0.132741786 0.08902115 0.003266755  0.0489595357
#>    16    3 0.98990998 0.592035104 0.85713494 0.581539292  0.9244150736
#>    69   78 0.89742362 0.331255674 0.83766611 0.524962424  0.9588264974
#>    17   81 0.71037641 0.147299428 0.65736202 0.290457923  0.6968440473
#>    86   76 0.67708195 0.113028631 0.62356571 0.095798510  0.6350435327
#>    76   99 0.62356571 0.095798510 0.50442667 0.134005351  0.5745511593
#>    99   28 0.50442667 0.134005351 0.46552480 0.012149823  0.4144151845
#>    28   37 0.46552480 0.012149823 0.08902115 0.003266755  0.3376240874
#>    86   77 0.67708195 0.113028631 0.61617783 0.205913084  0.6567396280
#>    14   51 0.78076938 0.213758161 0.74006973 0.332167096  0.8141938913
#>     4   26 0.85105617 0.011234569 0.69086843 0.036936092  0.7796369417
#>    14    4 0.78076938 0.213758161 0.85105617 0.011234569  0.8113108668
#>    76   26 0.62356571 0.095798510 0.69086843 0.036936092  0.5829379472
#>    14   54 0.78076938 0.213758161 0.94257511 0.112121343  0.8339100907
#>     4   28 0.85105617 0.011234569 0.46552480 0.012149823  0.6566281890
#>     4   47 0.85105617 0.011234569 0.02831269 0.000000000  0.6259011392
#>    79    2 0.17917927 0.798069152 0.15784677 0.721725445  0.1897820744
#>    97   83 0.15299606 0.825541117 0.06903141 0.872834801  0.0926973739
#>    42   50 0.12262931 0.685592500 0.22105670 0.721083940  0.1887913654
#>    93   83 0.00107170 0.952117394 0.06903141 0.872834801 -0.0137877913
#>    74   21 0.57994556 0.753167793 0.45884640 0.893689086  0.5085858748
#>    22   35 0.27666142 0.541416004 0.28951936 0.621634572  0.3041305600
#>    60   61 0.14134759 0.551806441 0.02704967 0.510602696  0.0750437695
#>    31   73 0.51017101 0.683175399 0.41467106 0.815877706  0.4900175402
#>    38   71 0.48556525 0.513867738 0.49436214 0.667672004  0.4029320499
#>    12   41 0.51449948 0.999085547 0.28215426 1.000000000  0.3983013154
#>    87   25 0.00000000 0.362680383 0.23130586 0.366725564  0.1155675741
#>    67   19 0.62846196 0.641495717 0.60596813 0.775706170  0.6382617653
#>   100   12 0.87166482 0.972873474 0.51449948 0.999085547  0.7078736667
#>    64   67 0.66328201 0.513660998 0.62846196 0.641495717  0.5890987319
#>    18   88 0.50461988 0.331818110 0.53416589 0.492592609  0.5509074912
#>    30   32 0.27094723 0.332618012 0.34266171 0.224802441  0.2761720067
#>    40   45 0.76140231 0.853897987 0.78505026 0.647601454  0.7590124695
#>    55   75 0.29818312 0.153175115 0.33756237 0.183582125  0.3278925341
#>    66   27 0.30211779 0.170344573 0.22540391 0.132741786  0.2535142886
#>    95    3 0.85135158 0.658832758 0.85713494 0.581539292  0.8736817170
#>    65   57 0.80918163 0.519524860 0.77017137 0.363658302  0.8412857570
#>    90   16 0.91203957 0.641903157 0.98990998 0.592035104  0.9889595323
#>    24  100 1.00000000 0.931993500 0.87166482 0.972873474  0.9240664282
#>    16   24 0.98990998 0.592035104 1.00000000 0.931993500  1.9563262988
#>    17   56 0.71037641 0.147299428 0.72867731 0.294424558  0.6968440473
#>    57   14 0.77017137 0.363658302 0.78076938 0.213758161  0.8196331703
#>    17   86 0.71037641 0.147299428 0.67708195 0.113028631  0.7400680585
#>    54   16 0.94257511 0.112121343 0.98990998 0.592035104  1.3629302154
#>           my1           mx2          my2 bp1 bp2
#>    0.82077836  0.2072650261   0.85105080   0   0
#>    0.75395415  0.1887913654   0.65633583   0   0
#>    0.75850905  0.1755025594   0.82077836   0   0
#>    0.76720333  0.1734812028   0.75850905   0   0
#>    0.83367769  0.2908148209   0.85229360   0   0
#>    0.81666931 -0.0457933092   0.82655345   0   0
#>    0.90031051  0.2072650261   0.85105080   0   0
#>    0.78245148  0.1755025594   0.82077836   0   0
#>    0.73642032  0.1202415620   0.76720333   0   0
#>    0.78538393  0.2627800227   0.74900275   0   0
#>    0.66502486  0.3704484955   0.66337719   0   0
#>    0.96720156  0.2908148209   0.85229360   0   0
#>    0.71672101  0.1066250558   0.73642032   0   0
#>    0.87061183  0.1279554990   0.94624951   0   0
#>    0.83142948 -3.1076046778   0.66305003   0   0
#>    0.61242065  0.0813028594   0.71672101   0   0
#>    0.75000852  0.0813028594   0.71672101   0   0
#>    1.03549215  0.1279554990   0.94624951   0   0
#>    0.82655345  0.0291142735   0.75000852   0   0
#>    1.16359086  0.1416038710   1.03549215   0   0
#>    0.78790370  0.4014136678   0.66066187   0   0
#>    0.80474550  0.3462914267   0.71470068   0   0
#>    0.71470068  0.2627800227   0.74900275   0   0
#>    0.66337719  0.3462914267   0.71470068   0   0
#>    0.64293987  0.1843842580   0.62603027   0   0
#>    0.66331611  0.3041305600   0.57815284   0   0
#>    0.61303777  0.0291142735   0.75000852   0   0
#>    0.82281799  0.3237231923   0.85633389   0   0
#>    0.64786690  0.3704484955   0.66337719   0   0
#>    0.85633389  0.2908148209   0.85229360   0   0
#>    0.89931675  0.3237231923   0.85633389   0   0
#>    0.62603027  0.2104511822   0.56545120   0   0
#>    0.98785529  0.2326683627   0.96720156   0   0
#>    0.81411245  0.3669908750   0.89931675   0   0
#>    0.72064903  0.4014136678   0.66066187   0   0
#>    0.53000836  0.2078361047   0.53139525   0   0
#>    0.53139525  0.2104511822   0.56545120   0   0
#>    0.55659986 -0.0007225237   0.43924687   0   0
#>    0.60813302  0.1843842580   0.62603027   0   0
#>    0.52859046  0.2620640598   0.44103810   0   0
#>    0.76938653  0.3928447179   0.78790370   0   0
#>    0.59574768  0.3912185046   0.64786690   0   0
#>    0.48673922  0.0997835210   0.39875424   0   0
#>    0.77091867  0.5775703246   0.68576075   0   0
#>    0.99305067  0.3951753921   0.98785529   0   0
#>    0.94504751  0.7078736667   1.18752856   0   0
#>    0.81246109  0.3437897880   0.82281799   0   0
#>    0.87224774  0.5738437542   0.87596140   0   0
#>    0.83451551  0.6382617653   0.71212840   0   0
#>    0.57473591  0.3817273843   0.56571513   0   0
#>    0.58752114  0.4899341008   0.45769897   0   0
#>    0.56571513  0.3652172197   0.53000836   0   0
#>    0.43544703  0.2620640598   0.44103810   0   0
#>    0.62166517  0.5775703246   0.68576075   0   0
#>    0.44103810  0.1155675741   0.36958361   0   0
#>    0.40640431  0.3396688931   0.45128296   0   0
#>    0.45128296  0.3652172197   0.53000836   0   0
#>    0.58789159  0.5890987319   0.56211427   0   0
#>    0.45868315  0.3839325402   0.40640431   0   0
#>    0.25252696  0.0476642345   0.18395599   0   0
#>    0.43924687 -0.7078896585   0.56856229   0   0
#>    0.56856229 -0.0007393129   0.61303777   0   0
#>    0.37947533  0.2961328304   0.34489323   0   0
#>    0.45769897  0.5509074912   0.40641383   0   0
#>    0.35816906  0.2961328304   0.34489323   0   0
#>    0.34489323  0.3020228340   0.29000646   0   0
#>    0.83278654  0.7424051955   1.03857421   0   0
#>    0.72225724  0.6390729564   0.58407478   0   0
#>    0.35555492  0.3894229142   0.39007051   0   0
#>    0.39007051  0.3839325402   0.40640431   0   0
#>    0.33943050  0.4544781894   0.22375400   0   0
#>    1.03857421  0.8088646559   0.92442916   0   0
#>    0.82387377  0.5738437542   0.87596140   0   0
#>    0.29000646  0.2761720067   0.25833477   0   0
#>    0.30490122  0.3978913106   0.23771501   0   0
#>    0.56985825  0.7251325144   0.50086913   0   0
#>    0.58362848  0.6390729564   0.58407478   0   0
#>    0.25129433  0.2209492472   0.23888362   0   0
#>    0.72582861  0.6986959634   0.72225724   0   0
#>    0.40251921  0.5181509323   0.23289842   0   0
#>    0.60261718  0.7048820681   0.58362848   0   0
#>    0.41401849  0.6314249127   0.40729516   0   0
#>    0.21451790  0.3978913106   0.23771501   0   0
#>    0.25833477  0.3978913106   0.23771501   0   0
#>    0.90167893  0.6849057929   0.83278654   0   0
#>    0.74912038  0.8025072746   0.74586078   0   0
#>    0.62890206  0.7262068096   0.60261718   0   0
#>    0.40686417  0.6666399746   0.38367006   0   0
#>    0.74453740  0.6791216714   0.82387377   0   0
#>    0.23294125  0.4573370168   0.35555492   0   0
#>    0.20814056  0.3278925341   0.15540228   0   0
#>    0.49527209  0.7251325144   0.50086913   0   0
#>    0.24590032  0.2209492472   0.23888362   0   0
#>    0.19293163  0.3081962715   0.20814056   0   0
#>    0.17244731  0.2743309120   0.09830289   0   0
#>    0.67330349  0.8702295852   0.62138218   0   0
#>    0.60609110  0.7375006261   0.62890206   0   0
#>    0.49641916  0.7290124482   0.49527209   0   0
#>    0.54221020  0.7262068096   0.60261718   0   0
#>    0.74586078  0.8719491162   0.76499197   0   0
#>    0.56043428  0.7724947568   0.54221020   0   0
#>    0.60640542  0.8147983270   0.67330349   0   0
#>    0.40729516  0.6083350139   0.25798315   0   0
#>    0.23176217  0.6083350139   0.25798315   0   0
#>    0.60924624  0.8006821838   0.60640542   0   0
#>    1.15091195  0.5738437542   0.87596140   0   0
#>    0.92442916  0.7918067663   0.90167893   0   0
#>    0.88331594  0.8088646559   0.92442916   0   0
#>    0.35183145  0.6303425526   0.25131646   0   0
#>    0.62164048  0.9244150736   0.57549538   0   0
#>    0.55267898  0.8434761499   0.60924624   0   0
#>    0.55143752  0.7973825457   0.56043428   0   0
#>    0.38138770  0.8395967451   0.42909755   0   0
#>    0.42867482  0.8532079705   0.42368619   0   0
#>    0.42909755  0.7455408880   0.49641916   0   0
#>    0.22162695  0.6303425526   0.25131646   0   0
#>    0.11986404  0.2743309120   0.09830289   0   0
#>    0.79041343  1.1271454951   0.83005107   0   0
#>    0.52538610  0.8490624062   0.55267898   0   0
#>    0.09830289  0.2591033433   0.07728097   0   0
#>    0.29025143  0.6911337595   0.32634756   0   0
#>    0.11329135  0.4144151845   0.09560375   0   0
#>    0.72585076  0.9889595323   0.67628343   0   0
#>    0.06555295  0.1968128302  -0.22774051   0   0
#>    0.12394636  0.1968128302  -0.22774051   0   0
#>    0.76499197  0.9027523518   0.72585076   0   0
#>    0.67628343  1.9563262988   0.73348066   0   0
#>   -0.03017414  0.2591033433   0.07728097   0   0
#>    0.18395599 -2.6928830733  -0.02998521   0   1
#>    0.91549635  0.8817545142   0.88331594   0   0
#>    0.18203276  0.4482171154  -7.23765866   0   0
#>    0.73348066  1.1271454951   0.83005107   0   0
#>    0.45626895  0.9283761918   0.52538610   0   0
#>    0.22368351  0.6911337595   0.32634756   0   0
#>    0.15187371  0.5745511593   0.14781512   0   0
#>    0.14781512  0.5205343427   0.23176217   0   0
#>    0.09560375  0.3701532561  -0.03017414   0   0
#>    0.29183056  0.8953896312   0.21661780   0   0
#>    0.05407777  0.5829379472  -0.01856273   0   0
#>   -2.55023810  0.1968128302  -0.22774051   0   0
#>    0.22255794  0.8113108668   0.11089925   0   0
#>    0.08514529  0.6611320260   0.07084363   0   0
#>    0.16609979  0.6912905756   0.22162695   0   0
#>    0.29144600  0.7201002567   0.38138770   0   0
#>    0.07815104  0.7400680585   0.08514529   0   0
#>    0.11089925  0.7796369417   0.07815104   0   0
#>   -0.01856273  0.6566281890  -0.68851538   0   0
#>    0.31295211  0.9588264974   0.45626895   0   0
#>    0.21661780  1.3629302154   0.31295211   0   0
#>    0.11874242  6.2714193979  -4.81386742   0   1
#>  -13.63162866  0.4482171154  -7.23765866   0   0
#>    0.85105080  0.2617317321   0.83367769   0   0
#>    0.75395415  0.2475608968   0.78538393   0   0
#>    0.76720333  0.1154991339   0.78245148   0   0
#>    0.81666931  0.1154991339   0.78245148   0   0
#>    0.90031051  0.1279554990   0.94624951   0   0
#>    0.65633583  0.1066250558   0.73642032   0   0
#>    0.78538393  0.2617317321   0.83367769   0   0
#>    0.74900275  0.2460866386   0.66502486   0   0
#>    0.87061183  0.1096667453   1.16359086   0   0
#>    0.83142948 -0.0457933092   0.82655345   0   0
#>    0.61242065 -0.0007393129   0.61303777   0   0
#>    1.03549215  0.2326683627   0.96720156   0   0
#>    1.16359086 -0.4792941288   4.62093460   0   1
#>    0.80474550  0.3928447179   0.78790370   0   0
#>    0.64293987  0.2436045029   0.66331611   0   0
#>    0.66331611  0.2460866386   0.66502486   0   0
#>    0.82281799  0.3537780368   0.80474550   0   0
#>    0.98785529  0.3669908750   0.89931675   0   0
#>    0.81411245  0.5079089234   0.81246109   0   0
#>    0.72064903  0.5549863217   0.62166517   0   0
#>    0.66066187  0.3912185046   0.64786690   0   0
#>    0.57815284  0.2104511822   0.56545120   0   0
#>    0.53139525  0.2071792918   0.52859046   0   0
#>    0.55659986  0.1207482783   0.48673922   0   0
#>    0.60813302  0.0486164667   0.61242065   0   0
#>    0.76938653  0.4579151357   0.72064903   0   0
#>    0.59574768  0.3913818384   0.57473591   0   0
#>    0.48673922  0.2071792918   0.52859046   0   0
#>    0.77091867  0.5079089234   0.81246109   0   0
#>    0.99305067  0.4892102866   0.94504751   0   0
#>    0.87224774  0.4892102866   0.94504751   0   0
#>    0.83451551  0.5625202275   0.87224774   0   0
#>    0.57473591  0.4852952692   0.45868315   0   0
#>    0.58752114  0.5483937181   0.58789159   0   0
#>    0.43544703  0.2767695970   0.37947533   0   0
#>    0.36958361  0.0997835210   0.39875424   0   0
#>    0.39875424 -0.0007225237   0.43924687   0   0
#>    0.45128296  0.2810570421   0.43544703   0   0
#>    0.58789159  0.5549863217   0.62166517   0   0
#>    0.45868315  0.4899341008   0.45769897   0   0
#>    0.71212840  0.5775703246   0.68576075   0   0
#>    0.25252696  0.1664823032   0.25129433   0   0
#>    0.56856229 -3.1076046778   0.66305003   0   0
#>    0.66305003 -4.8041331899   0.66613462   0   1
#>    0.35816906  0.3728189413   0.33943050   0   0
#>    1.18752856  0.7075536168   1.15091195   0   0
#>    0.83278654  0.6791216714   0.82387377   0   0
#>    0.56211427  0.6132641073   0.41401849   0   0
#>    0.40641383  0.4573370168   0.35555492   0   0
#>    0.39007051  0.3764448996   0.35816906   0   0
#>    1.03857421  0.7075536168   1.15091195   0   0
#>    0.29000646  0.3412316533   0.30490122   0   0
#>    0.30490122  0.3728189413   0.33943050   0   0
#>    0.56985825  0.6390729564   0.58407478   0   0
#>    0.58362848  0.7251325144   0.50086913   0   0
#>    0.25129433  0.2767695970   0.37947533   0   0
#>    0.72582861  0.7375006261   0.62890206   0   0
#>    0.40251921  0.6132641073   0.41401849   0   0
#>    0.22375400  0.4432533374   0.21451790   0   0
#>    0.21451790  0.4311371736   0.19293163   0   0
#>    0.90167893  0.7590124695   0.74912038   0   0
#>    0.40686417  0.7290124482   0.49527209   0   0
#>    0.74453740  0.7056490056   0.72582861   0   0
#>    0.23294125  0.4544781894   0.22375400   0   0
#>    0.23289842  0.4742989791   0.23294125   0   0
#>    0.24590032  0.3081962715   0.20814056   0   0
#>    0.19293163  0.4074753822   0.11329135   0   0
#>    0.15540228  0.2535142886   0.17244731   0   0
#>    0.67330349  0.8025072746   0.74586078   0   0
#>    0.60609110  0.7973825457   0.56043428   0   0
#>    0.54221020  0.7455408880   0.49641916   0   0
#>    0.60640542  0.8002572862   0.60609110   0   0
#>    0.40729516  0.6362080086   0.40686417   0   0
#>    0.23176217  0.5181509323   0.23289842   0   0
#>    0.88331594  0.8840281777   0.79041343   0   0
#>    0.35183145  0.6911337595   0.32634756   0   0
#>    0.38367006  0.6782823338   0.35183145   0   0
#>    0.25798315  0.6303425526   0.25131646   0   0
#>    0.62164048  0.9027523518   0.72585076   0   0
#>    0.62138218  0.8434761499   0.60924624   0   0
#>    0.55267898  0.8178508881   0.55143752   0   0
#>    0.55143752  0.8412857570   0.42867482   0   0
#>    0.38138770  0.6666399746   0.38367006   0   0
#>    0.11986404  0.4074753822   0.11329135   0   0
#>    0.42368619  0.8196331703   0.29183056   0   0
#>    0.29025143  0.7058927427   0.22255794   0   0
#>    0.06555295  0.2591033433   0.07728097   0   0
#>    0.12394636  0.2316668254   0.06555295   0   0
#>    0.76499197  0.8840281777   0.79041343   0   0
#>   -0.03017414  0.3376240874  -2.55023810   0   0
#>    0.18395599  0.0489595357   0.18203276   0   0
#>    0.91549635  1.1271454951   0.83005107   0   0
#>    0.18203276  0.1041038873   0.12394636   0   0
#>    0.57549538  0.9283761918   0.52538610   0   0
#>    0.45626895  0.8532079705   0.42368619   0   0
#>    0.22368351  0.6912905756   0.22162695   0   0
#>    0.15187371  0.6611320260   0.07084363   0   0
#>    0.14781512  0.5444904162   0.05407777   0   0
#>    0.09560375  0.5444904162   0.05407777   0   0
#>   -2.55023810  0.4482171154  -7.23765866   0   0
#>    0.16609979  0.6350435327   0.15187371   0   0
#>    0.29144600  0.8107184812   0.29025143   0   0
#>    0.07815104  0.6566281890  -0.68851538   0   0
#>    0.11089925  0.8339100907   0.11874242   0   0
#>   -0.01856273  0.6611320260   0.07084363   0   0
#>    0.11874242  0.8953896312   0.21661780   0   0
#>   -0.68851538  0.6259011392 -13.63162866   0   0
#>  -13.63162866  0.6424949657 -14.84684774   0   1
#>    0.75395415  0.1734812028   0.75850905   0   0
#>    0.81666931  0.1398088730   0.90031051   0   0
#>    0.65633583  0.1936217468   0.64293987   0   0
#>    0.87061183 -0.0594983849   0.83142948   0   0
#>    0.81411245  0.5322612210   0.83451551   0   0
#>    0.57815284  0.3817273843   0.56571513   0   0
#>    0.55659986  0.0564663606   0.60813302   0   0
#>    0.76938653  0.4921465294   0.77091867   0   0
#>    0.59574768  0.5467642853   0.58752114   0   0
#>    0.99305067  0.4152402896   5.29692451   0   1
#>    0.36958361  0.1176147135   0.25252696   0   0
#>    0.71212840  0.6986959634   0.72225724   0   0
#>    1.18752856  0.9122496833   3.97235330   0   1
#>    0.56211427  0.6175291749   0.56985825   0   0
#>    0.40641383  0.5721000534   0.40251921   0   0
#>    0.25833477  0.2574780402   0.24590032   0   0
#>    0.74912038  0.7190321629   0.74453740   0   0
#>    0.15540228  0.3553336745   0.11986404   0   0
#>    0.17244731  0.2209492472   0.23888362   0   0
#>    0.62164048  0.8702295852   0.62138218   0   0
#>    0.42867482  0.8395967451   0.42909755   0   0
#>    0.67628343  0.9244150736   0.57549538   0   0
#>    0.91549635  3.1775126666   7.98977758   0   1
#>    0.73348066  4.8952737620   0.64625225   0   1
#>    0.22368351  0.7058927427   0.22255794   0   0
#>    0.29183056  0.8141938913   0.29144600   0   0
#>    0.08514529  0.6567396280   0.16609979   0   0
#>    0.31295211  3.4265623292   0.10941189   0   1
#> 
#> $x
#>              [,1]        [,2]
#>   [1,] 0.17036079 0.798355414
#>   [2,] 0.15784677 0.721725445
#>   [3,] 0.85713494 0.581539292
#>   [4,] 0.85105617 0.011234569
#>   [5,] 0.41827390 0.319873297
#>   [6,] 0.32334952 0.338241401
#>   [7,] 0.17671769 0.026384631
#>   [8,] 0.97997541 0.760056758
#>   [9,] 0.38749032 0.752828567
#>  [10,] 0.37602833 0.290050978
#>  [11,] 0.43126797 0.469928725
#>  [12,] 0.51449948 0.999085547
#>  [13,] 0.96536426 0.849674222
#>  [14,] 0.78076938 0.213758161
#>  [15,] 0.40285363 0.753863191
#>  [16,] 0.98990998 0.592035104
#>  [17,] 0.71037641 0.147299428
#>  [18,] 0.50461988 0.331818110
#>  [19,] 0.60596813 0.775706170
#>  [20,] 0.31976105 0.359949877
#>  [21,] 0.45884640 0.893689086
#>  [22,] 0.27666142 0.541416004
#>  [23,] 0.18456754 0.462548785
#>  [24,] 1.00000000 0.931993500
#>  [25,] 0.23130586 0.366725564
#>  [26,] 0.69086843 0.036936092
#>  [27,] 0.22540391 0.132741786
#>  [28,] 0.46552480 0.012149823
#>  [29,] 0.32230567 0.062549420
#>  [30,] 0.27094723 0.332618012
#>  [31,] 0.51017101 0.683175399
#>  [32,] 0.34266171 0.224802441
#>  [33,] 0.41498122 0.820792646
#>  [34,] 0.05287777 0.493927366
#>  [35,] 0.28951936 0.621634572
#>  [36,] 0.02960219 0.714746005
#>  [37,] 0.08902115 0.003266755
#>  [38,] 0.48556525 0.513867738
#>  [39,] 0.54035541 0.975769306
#>  [40,] 0.76140231 0.853897987
#>  [41,] 0.28215426 1.000000000
#>  [42,] 0.12262931 0.685592500
#>  [43,] 0.73485449 0.563275863
#>  [44,] 0.31195484 0.759108825
#>  [45,] 0.78505026 0.647601454
#>  [46,] 0.27647413 0.520163133
#>  [47,] 0.02831269 0.000000000
#>  [48,] 0.83550071 0.579402820
#>  [49,] 0.06129480 0.764434083
#>  [50,] 0.22105670 0.721083940
#>  [51,] 0.74006973 0.332167096
#>  [52,] 0.68769353 0.551736014
#>  [53,] 0.79989055 0.845624476
#>  [54,] 0.94257511 0.112121343
#>  [55,] 0.29818312 0.153175115
#>  [56,] 0.72867731 0.294424558
#>  [57,] 0.77017137 0.363658302
#>  [58,] 0.29065366 0.707249242
#>  [59,] 0.28878545 0.947824670
#>  [60,] 0.14134759 0.551806441
#>  [61,] 0.02704967 0.510602696
#>  [62,] 0.03416569 0.938172356
#>  [63,] 0.68812767 0.615751138
#>  [64,] 0.66328201 0.513660998
#>  [65,] 0.80918163 0.519524860
#>  [66,] 0.30211779 0.170344573
#>  [67,] 0.62846196 0.641495717
#>  [68,] 0.46313464 0.408142881
#>  [69,] 0.89742362 0.331255674
#>  [70,] 0.27590595 0.014597437
#>  [71,] 0.49436214 0.667672004
#>  [72,] 0.01399636 0.718147388
#>  [73,] 0.41467106 0.815877706
#>  [74,] 0.57994556 0.753167793
#>  [75,] 0.33756237 0.183582125
#>  [76,] 0.62356571 0.095798510
#>  [77,] 0.61617783 0.205913084
#>  [78,] 0.83766611 0.524962424
#>  [79,] 0.17917927 0.798069152
#>  [80,] 0.76194467 0.584034209
#>  [81,] 0.65736202 0.290457923
#>  [82,] 0.83162585 0.637773550
#>  [83,] 0.06903141 0.872834801
#>  [84,] 0.73573316 0.893729048
#>  [85,] 0.21504412 0.910509614
#>  [86,] 0.67708195 0.113028631
#>  [87,] 0.00000000 0.362680383
#>  [88,] 0.53416589 0.492592609
#>  [89,] 0.74419213 0.428727221
#>  [90,] 0.91203957 0.641903157
#>  [91,] 0.64375476 0.296947795
#>  [92,] 0.67302977 0.911249084
#>  [93,] 0.00107170 0.952117394
#>  [94,] 0.34999605 0.268097367
#>  [95,] 0.85135158 0.658832758
#>  [96,] 0.58633945 0.305826610
#>  [97,] 0.15299606 0.825541117
#>  [98,] 0.26787097 0.606248929
#>  [99,] 0.50442667 0.134005351
#> [100,] 0.87166482 0.972873474
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,79,97), [2,9,4]
#> 2: (85,97,79), [1,6,8]
#> 3: (79,2,50), [10,12,4]
#> 4: (2,79,1), [1,5,3]
#> 5: (49,2,1), [4,9,11]
#> 6: (79,44,85), [16,2,12]
#> 7: (97,83,49), [24,9,8]
#> 8: (83,97,85), [2,20,7]
#> 9: (49,1,97), [1,7,5]
#> 10: (42,50,2), [3,11,30]
#> 11: (42,2,49), [5,17,10]
#> 12: (79,50,44), [13,6,3]
#> 13: (58,44,50), [12,14,28]
#> 14: (50,35,58), [29,13,31]
#> 15: (41,85,59), [16,38,23]
#> 16: (59,85,44), [6,35,15]
#> 17: (36,42,49), [11,22,21]
#> 18: (93,83,62), [20,25,19]
#> 19: (83,93,72), [75,24,18]
#> 20: (85,62,83), [18,8,23]
#> 21: (61,42,36), [17,32,47]
#> 22: (72,36,49), [17,24,32]
#> 23: (41,62,85), [20,15,25]
#> 24: (83,72,49), [22,7,19]
#> 25: (93,62,41), [23,0,18]
#> 26: (73,9,15), [41,49,27]
#> 27: (73,44,9), [28,26,33]
#> 28: (9,44,58), [13,29,27]
#> 29: (35,9,58), [28,14,34]
#> 30: (50,42,98), [37,31,10]
#> 31: (50,98,35), [42,14,30]
#> 32: (61,36,72), [22,74,21]
#> 33: (73,33,44), [35,27,55]
#> 34: (71,9,35), [29,50,41]
#> 35: (33,59,44), [16,33,36]
#> 36: (21,59,33), [35,39,38]
#> 37: (42,60,98), [43,30,47]
#> 38: (21,41,59), [15,36,53]
#> 39: (74,21,33), [36,55,57]
#> 40: (31,15,71), [41,63,49]
#> 41: (71,15,9), [26,34,40]
#> 42: (22,35,98), [31,43,61]
#> 43: (22,98,60), [37,45,42]
#> 44: (11,22,46), [45,68,61]
#> 45: (46,22,60), [43,48,44]
#> 46: (60,61,34), [73,51,47]
#> 47: (61,60,42), [37,21,46]
#> 48: (60,23,46), [64,45,51]
#> 49: (31,73,15), [26,40,52]
#> 50: (38,71,35), [34,59,60]
#> 51: (60,34,23), [66,48,46]
#> 52: (73,31,74), [58,55,49]
#> 53: (12,41,21), [38,54,0]
#> 54: (21,39,12), [80,53,56]
#> 55: (74,33,73), [33,52,39]
#> 56: (21,19,39), [81,54,57]
#> 57: (21,74,19), [71,56,39]
#> 58: (67,74,31), [52,63,71]
#> 59: (38,35,11), [61,70,50]
#> 60: (71,38,88), [77,69,50]
#> 61: (35,22,11), [44,59,42]
#> 62: (20,46,25), [64,76,68]
#> 63: (71,67,31), [58,40,69]
#> 64: (46,23,25), [65,62,48]
#> 65: (87,25,23), [64,66,72]
#> 66: (87,23,34), [51,73,65]
#> 67: (68,11,20), [68,87,70]
#> 68: (20,11,46), [44,62,67]
#> 69: (71,88,67), [84,63,60]
#> 70: (38,11,68), [67,77,59]
#> 71: (67,19,74), [57,58,83]
#> 72: (25,87,27), [161,97,65]
#> 73: (34,61,87), [74,66,46]
#> 74: (87,61,72), [32,75,73]
#> 75: (87,72,93), [19,0,74]
#> 76: (25,30,20), [79,62,97]
#> 77: (38,68,88), [86,60,70]
#> 78: (5,20,6), [79,88,87]
#> 79: (20,30,6), [91,78,76]
#> 80: (100,12,39), [54,130,0]
#> 81: (92,39,19), [56,90,130]
#> 82: (40,84,92), [89,90,106]
#> 83: (19,67,63), [95,98,71]
#> 84: (64,67,88), [69,103,93]
#> 85: (18,68,5), [87,111,86]
#> 86: (18,88,68), [77,85,99]
#> 87: (5,68,20), [67,78,85]
#> 88: (6,10,5), [100,78,92]
#> 89: (92,84,100), [131,130,82]
#> 90: (40,92,19), [81,110,82]
#> 91: (6,30,94), [105,92,79]
#> 92: (6,94,10), [96,88,91]
#> 93: (67,64,52), [102,95,84]
#> 94: (43,63,52), [95,102,101]
#> 95: (67,52,63), [94,83,93]
#> 96: (32,10,94), [92,105,104]
#> 97: (25,27,30), [122,76,72]
#> 98: (45,19,63), [83,108,110]
#> 99: (88,18,96), [114,103,86]
#> 100: (99,5,10), [88,104,111]
#> 101: (80,63,43), [94,123,108]
#> 102: (43,52,64), [93,113,94]
#> 103: (88,96,64), [127,84,99]
#> 104: (99,10,32), [96,116,100]
#> 105: (30,32,94), [96,91,115]
#> 106: (53,84,40), [82,107,131]
#> 107: (40,45,53), [124,106,110]
#> 108: (45,63,80), [101,120,98]
#> 109: (64,91,89), [134,113,127]
#> 110: (45,40,19), [90,98,107]
#> 111: (99,18,5), [85,100,114]
#> 112: (32,66,75), [117,116,115]
#> 113: (89,43,64), [102,109,121]
#> 114: (99,96,18), [99,111,128]
#> 115: (32,30,66), [122,112,105]
#> 116: (99,32,75), [112,152,104]
#> 117: (55,75,66), [112,118,145]
#> 118: (66,27,55), [149,117,122]
#> 119: (45,82,95), [137,124,126]
#> 120: (48,45,80), [108,125,126]
#> 121: (65,43,89), [113,142,123]
#> 122: (27,66,30), [115,97,118]
#> 123: (65,80,43), [101,121,125]
#> 124: (45,95,53), [156,107,119]
#> 125: (48,80,65), [123,139,120]
#> 126: (48,82,45), [119,120,129]
#> 127: (64,96,91), [135,109,103]
#> 128: (99,77,96), [135,114,170]
#> 129: (3,82,48), [126,138,137]
#> 130: (100,39,92), [81,89,80]
#> 131: (100,84,53), [106,132,89]
#> 132: (13,100,53), [131,146,162]
#> 133: (51,91,81), [144,151,134]
#> 134: (51,89,91), [109,133,140]
#> 135: (91,96,77), [128,144,127]
#> 136: (95,3,90), [166,153,137]
#> 137: (3,95,82), [119,129,136]
#> 138: (78,3,48), [129,139,147]
#> 139: (78,48,65), [125,141,138]
#> 140: (51,57,89), [142,134,180]
#> 141: (65,57,78), [148,139,142]
#> 142: (57,65,89), [121,140,141]
#> 143: (17,81,77), [144,178,168]
#> 144: (91,77,81), [143,133,135]
#> 145: (75,55,29), [149,152,117]
#> 146: (53,8,13), [163,132,156]
#> 147: (16,3,78), [138,167,166]
#> 148: (69,78,57), [141,172,167]
#> 149: (55,27,29), [159,145,118]
#> 150: (14,51,56), [151,176,180]
#> 151: (81,56,51), [150,133,168]
#> 152: (75,29,99), [171,116,145]
#> 153: (95,90,8), [157,156,136]
#> 154: (27,7,70), [160,159,155]
#> 155: (27,37,7), [160,154,164]
#> 156: (53,95,8), [153,146,124]
#> 157: (90,16,8), [165,153,166]
#> 158: (28,29,70), [159,174,171]
#> 159: (27,70,29), [158,149,154]
#> 160: (37,70,7), [154,155,174]
#> 161: (27,87,47), [0,164,72]
#> 162: (24,100,13), [132,163,0]
#> 163: (24,13,8), [146,165,162]
#> 164: (27,47,37), [175,155,161]
#> 165: (16,24,8), [163,157,0]
#> 166: (16,90,3), [136,147,157]
#> 167: (69,16,78), [147,148,184]
#> 168: (17,56,81), [151,143,176]
#> 169: (86,77,76), [170,179,178]
#> 170: (76,77,99), [128,173,169]
#> 171: (99,29,28), [158,173,152]
#> 172: (57,14,69), [185,148,180]
#> 173: (99,28,76), [183,170,171]
#> 174: (28,70,37), [160,175,158]
#> 175: (28,37,47), [164,188,174]
#> 176: (56,17,14), [182,150,168]
#> 177: (17,86,26), [179,181,178]
#> 178: (86,17,77), [143,169,177]
#> 179: (76,26,86), [177,169,183]
#> 180: (14,57,51), [140,150,172]
#> 181: (4,17,26), [177,187,182]
#> 182: (14,17,4), [181,186,176]
#> 183: (76,28,26), [187,179,173]
#> 184: (54,16,69), [167,185,0]
#> 185: (14,54,69), [184,172,186]
#> 186: (14,4,54), [0,185,182]
#> 187: (4,26,28), [183,188,181]
#> 188: (4,28,47), [175,0,187]
#> boundary nodes:  41 93 87 47 4 54 16 24 100 12 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.037950882 0.063213189 0.008823128 0.077645043 0.179760638 0.108676451 
#>                                                                         
#> 0.105212201 0.032258275 0.105576088 0.098531027 0.085622185 0.082645005 
#>                                                                         
#> 0.099889528 0.074058181 0.234326718 0.097488321 0.058934936 0.182981507 
#>                                                                         
#> 0.066178551 0.255579704 0.015398109 0.075796107 0.056063861 0.107027158 
#>                                                                         
#> 0.165500876 0.026558819 0.015972204 0.120080509 0.163738125 0.190132825 
#>                                                                         
#> 0.179060070 0.137739427 0.052595035 0.085076699 0.125708902 0.021253697 
#>                                                                         
#> 0.135712174 0.030743417 0.135089168 0.108472278 0.063130433 0.209952399 
#>                                                                         
#> 0.135376546 0.098830275 0.206208931 0.034816327 0.004924717 0.210547589 
#>                                                                         
#> 0.034426047 0.207623055 0.053053307 0.170333839 0.159947680 0.125419079 
#>                                                                         
#> 0.106614072 0.156617818 0.162740984 0.176249495 0.069519602 0.321991743 
#>                                                                         
#> 0.150375181 0.207954774 0.055944788 0.110350333 0.022003070 0.052703155 
#>                                                                         
#> 0.065105043 0.064982924 0.099015177 0.151256575 0.051711303 0.157293532 
#>                                                                         
#> 0.151225478 0.102037391 0.034053494 0.045228634 0.064016597 0.204999267 
#>                                                                         
#> 0.179821584 0.085753397 0.070264143 0.221619699 0.073285079 0.043911766 
#>                                                                         
#> 0.047385849 0.198578326 0.080342471 0.165690927 0.173993517 0.087168272 
#>                                                                         
#> 0.037835846 0.087087626 0.165240045 0.041534535 0.075593211 0.028854711 
#>                                                                         
#> 0.067636254 0.134872268 0.034128974 0.193750827 0.079954904 0.047601194 
#>                                                                         
#> 0.058097772 0.104273880 0.058499202 0.147530825 0.080188656 0.146095355 
#>                                                                         
#> 0.015075682 0.081598497 0.021739468 0.065406908 0.070063437 0.174855827 
#>                                                                         
#> 0.111659512 0.094042369 0.093781205 0.090800748 0.059832909 0.119653330 
#>                                                                         
#> 0.039424441 0.195637440 0.136292159 0.099886172 0.090692445 0.163677702 
#>                                                                         
#> 0.168315096 0.066725761 0.363783822 0.154782561 0.060796289 0.173098895 
#>                                                                         
#> 0.166363879 0.071425515 0.110362131 0.132887337 0.151828316 0.165571367 
#>                                                                         
#> 0.178812837 0.187227965 0.096808748 0.077331374 0.110945626 0.043563801 
#>                                                                         
#> 0.112074204 0.195715208 0.226702690 0.276694124 0.223737597 0.136212546 
#>                                                                         
#> 0.437380891 0.118021798 0.087638132 0.114219291 0.110196145 0.150794904 
#>                                                                         
#> 0.050456505 0.138373621 0.070958690 0.035912066 0.164186030 0.204159266 
#>                                                                         
#> 0.111858392 0.285131808 0.068658477 0.124010705 0.120736540 0.117359819 
#>                                                                         
#> 0.178469573 0.178287279 0.022142170 0.136650014 0.065426143 0.138782138 
#>                                                                         
#> 0.105720837 0.199391325 0.128505970 0.223713393 0.099171030 0.176771587 
#>                                                                         
#> 0.119187580 0.115675763 0.188586249 0.069848632 0.179547022 0.088714320 
#>                                                                         
#> 0.209854410 0.141498514 0.165957960 0.136630747 0.108078110 0.121755878 
#>                                                                         
#> 0.234058201 0.355742448 0.589437985 0.096685179 0.331322059 0.105351099 
#>                                                                         
#> 0.130823733 0.086870573 0.106352780 0.207974629 0.075034811 0.071395890 
#>                                                                         
#> 0.107541543 0.048552278 0.052294969 0.102021714 0.193916524 0.202080113 
#>                                                                         
#> 0.185504756 0.039367444 0.117303854 0.220184527 0.204863841 0.197812853 
#>                                                                         
#> 0.067893068 0.174073406 0.017614537 0.067245868 0.073701696 0.086247751 
#>                                                                         
#> 0.084831016 0.217591190 0.190347702 0.165523257 0.092629478 0.102552298 
#>                                                                         
#> 0.095119952 0.063005110 0.061749536 0.054483444 0.028998841 0.096648083 
#>                                                                         
#> 0.121990502 0.131312854 0.096024223 0.128485573 0.116970904 0.199380020 
#>                                                                         
#> 0.189634649 0.237624348 0.089309003 0.188053293 0.133189246 0.202714737 
#>                                                                         
#> 0.152659358 0.056221572 0.125115436 0.127914523 0.376608424 0.111071295 
#>                                                                         
#> 0.125208376 0.162236501 0.214373596 0.089411636 0.191078880 0.385532460 
#>                                                                         
#> 0.822820181 0.079268135 0.096367812 0.104630747 0.104423424 0.185502670 
#>                                                                         
#> 0.081242508 0.121497997 0.163493556 0.154055631 0.232347019 0.231341227 
#>                                                                         
#> 0.136082394 0.358125891 0.132492079 0.163466836 0.129488081 0.207647501 
#>                                                                         
#> 0.049752507 0.085434117 0.077509529 0.160674156 0.092469585 0.134688869 
#>                                                             
#> 0.340108099 0.148258985 0.150274316 0.047780839 0.482242479 
#> 
#> $alpha
#> [1] 0.3124913
#> 
#> attr(,"class")
#> [1] "scree"