Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1           y1          x2           y2          mx1
#>    51    9 0.008634808 0.2419158597 0.136210111 0.3325412972   0.06199801
#>     9   21 0.136210111 0.3325412972 0.073139709 0.4182701574   0.10438641
#>    21   43 0.073139709 0.4182701574 0.041886486 0.4371951477   0.04634345
#>    43   45 0.041886486 0.4371951477 0.003881015 0.4696155084  -0.02536169
#>    45   51 0.003881015 0.4696155084 0.008634808 0.2419158597  -0.10561965
#>    87   74 0.671945315 0.0437160870 0.774725699 0.1178986432   0.68866553
#>    74   13 0.774725699 0.1178986432 0.813898244 0.1983189485   0.71477716
#>    13   29 0.813898244 0.1983189485 0.637692355 0.3635530003   0.70813191
#>    29   37 0.637692355 0.3635530003 0.582485175 0.2349416669   0.69268561
#>    37   86 0.582485175 0.2349416669 0.505427515 0.2218324095   0.55262078
#>    86   92 0.505427515 0.2218324095 0.442930085 0.0897716329   0.53332228
#>    92    4 0.442930085 0.0897716329 0.456812984 0.0634587788   0.53619154
#>     4    3 0.456812984 0.0634587788 0.552175139 0.0239003649   0.53673416
#>     3   87 0.552175139 0.0239003649 0.671945315 0.0437160870   0.60213916
#>    76   88 0.319470241 0.0288505019 0.057750178 0.0000000000   0.39562109
#>    88   71 0.057750178 0.0000000000 0.795106635 0.0004359111   0.42792527
#>    71   40 0.795106635 0.0004359111 0.681298621 0.0378172360   0.64769894
#>    40   87 0.681298621 0.0378172360 0.671945315 0.0437160870   0.62456533
#>     4   10 0.456812984 0.0634587788 0.355200004 0.0473170282   0.43462324
#>    10   76 0.355200004 0.0473170282 0.319470241 0.0288505019   0.43216858
#>    92   10 0.442930085 0.0897716329 0.355200004 0.0473170282   0.40628737
#>    25  100 0.884801800 0.4494671457 0.891743805 0.5325122165   0.79155383
#>   100   50 0.891743805 0.5325122165 0.679391963 0.6363686882   0.79390409
#>    50   36 0.679391963 0.6363686882 0.595882574 0.5901742426   0.65471158
#>    36   83 0.595882574 0.5901742426 0.674323558 0.4811310187   0.62529010
#>    83   22 0.674323558 0.4811310187 0.719541862 0.4140303035   0.75358163
#>    22   25 0.719541862 0.4140303035 0.884801800 0.4494671457   0.78895507
#>    64   44 0.340082972 0.9565561379 0.095749352 1.0000000000   0.26351058
#>    44   26 0.095749352 1.0000000000 0.170511389 0.9137596981   0.15488999
#>    26   81 0.170511389 0.9137596981 0.215156565 0.9441810056   0.18506149
#>    81   39 0.215156565 0.9441810056 0.333064705 0.9350749527   0.27583885
#>    39   64 0.333064705 0.9350749527 0.340082972 0.9565561379   0.27629306
#>    86   79 0.505427515 0.2218324095 0.457300490 0.2807282227   0.43542724
#>    79   16 0.457300490 0.2807282227 0.422266874 0.3600743151   0.38669491
#>    16   63 0.422266874 0.3600743151 0.233757248 0.3776044701   0.32379579
#>    63   15 0.233757248 0.3776044701 0.314740855 0.1659393325   0.25826560
#>    15   91 0.314740855 0.1659393325 0.353479724 0.0914889201   0.38809614
#>    91   92 0.353479724 0.0914889201 0.442930085 0.0897716329   0.39945476
#>    71   80 0.795106635 0.0004359111 0.987003000 0.0360723231   0.88700940
#>    80   69 0.987003000 0.0360723231 0.929489754 0.1481877889   0.90241877
#>    69   13 0.929489754 0.1481877889 0.813898244 0.1983189485   0.87693614
#>    74   71 0.774725699 0.1178986432 0.795106635 0.0004359111   0.81973716
#>    51   62 0.008634808 0.2419158597 0.046665551 0.1264752971   0.12324287
#>    62   15 0.046665551 0.1264752971 0.314740855 0.1659393325   0.17370276
#>    63   82 0.233757248 0.3776044701 0.185703254 0.4260788618   0.17793780
#>    82   21 0.185703254 0.4260788618 0.073139709 0.4182701574   0.13130671
#>    91   76 0.353479724 0.0914889201 0.319470241 0.0288505019   0.32180305
#>    99   68 0.668227860 0.8731021291 0.607236196 0.9833642085   0.61962175
#>    68   97 0.607236196 0.9833642085 0.542944764 0.9327711795   0.59969373
#>    97   33 0.542944764 0.9327711795 0.491173337 0.9222670774   0.52359461
#>    33   24 0.491173337 0.9222670774 0.464659455 0.8791219055   0.51429332
#>    24   60 0.464659455 0.8791219055 0.624188854 0.7796834205   0.51741602
#>    60   99 0.624188854 0.7796834205 0.668227860 0.8731021291   0.61279588
#>   100   38 0.891743805 0.5325122165 0.899256631 0.4782906157   0.93507612
#>    38   28 0.899256631 0.4782906157 0.976833765 0.4063398645   0.98376212
#>    28   66 0.976833765 0.4063398645 1.000000000 0.7079352223   1.15951586
#>    66   18 1.000000000 0.7079352223 0.966598831 0.6861169685   1.05031117
#>    18  100 0.966598831 0.6861169685 0.891743805 0.5325122165   0.83914313
#>    69   77 0.929489754 0.1481877889 0.943741881 0.2038872392   0.87926568
#>    77   54 0.943741881 0.2038872392 0.849510993 0.3516637626   0.87609349
#>    54   22 0.849510993 0.3516637626 0.719541862 0.4140303035   0.74269100
#>    22   29 0.719541862 0.4140303035 0.637692355 0.3635530003   0.73746075
#>    85   44 0.118735128 0.9116301637 0.095749352 1.0000000000   0.05085242
#>    44   17 0.095749352 1.0000000000 0.103577689 0.8501654308  -0.11518667
#>    17   26 0.103577689 0.8501654308 0.170511389 0.9137596981   0.15080059
#>    26   85 0.170511389 0.9137596981 0.118735128 0.9116301637   0.14559048
#>    62   48 0.046665551 0.1264752971 0.154210921 0.0461049700   0.17486295
#>    48   76 0.154210921 0.0461049700 0.319470241 0.0288505019   0.24282890
#>    79   53 0.457300490 0.2807282227 0.431394748 0.3592419102   0.42255721
#>    53   41 0.431394748 0.3592419102 0.418953220 0.5946782097   0.43758784
#>    41   20 0.418953220 0.5946782097 0.304105854 0.4415576258   0.41593648
#>    20   63 0.304105854 0.4415576258 0.233757248 0.3776044701   0.32596649
#>    44   61 0.095749352 1.0000000000 0.002896245 0.7026351919  -0.14583508
#>    61   57 0.002896245 0.7026351919 0.166177808 0.7320144601   0.07680107
#>    57   26 0.166177808 0.7320144601 0.170511389 0.9137596981   0.19374816
#>    66   72 1.000000000 0.7079352223 0.884158428 0.8450487322   0.93519184
#>    72   23 0.884158428 0.8450487322 0.848202332 0.8184849075   0.91793041
#>    23   50 0.848202332 0.8184849075 0.679391963 0.6363686882   0.82066146
#>    25   54 0.884801800 0.4494671457 0.849510993 0.3516637626   0.84715429
#>    54   32 0.849510993 0.3516637626 0.991126601 0.2234403178   0.93883956
#>    32   28 0.991126601 0.2234403178 0.976833765 0.4063398645   0.94683379
#>    28   25 0.976833765 0.4063398645 0.884801800 0.4494671457   0.93699043
#>    41   75 0.418953220 0.5946782097 0.307689589 0.6171812476   0.34766473
#>    75   27 0.307689589 0.6171812476 0.218677603 0.4644675100   0.28199966
#>    27   56 0.218677603 0.4644675100 0.240675041 0.4487273022   0.28194125
#>    56   63 0.240675041 0.4487273022 0.233757248 0.3776044701   0.26843289
#>    82   47 0.185703254 0.4260788618 0.129878271 0.4581370937   0.13124537
#>    47   35 0.129878271 0.4581370937 0.095636249 0.4614550545   0.10955538
#>    35   93 0.095636249 0.4614550545 0.073320062 0.4612126200   0.08471325
#>    93   43 0.073320062 0.4612126200 0.041886486 0.4371951477   0.06480626
#>    83   65 0.674323558 0.4811310187 0.647178053 0.4068854098   0.68077002
#>    65   29 0.647178053 0.4068854098 0.637692355 0.3635530003   0.68681104
#>    72   96 0.884158428 0.8450487322 0.830077798 0.8920302043   0.84733395
#>    96   99 0.830077798 0.8920302043 0.668227860 0.8731021291   0.75477791
#>    60   50 0.624188854 0.7796834205 0.679391963 0.6363686882   0.74570215
#>    33   64 0.491173337 0.9222670774 0.340082972 0.9565561379   0.41554066
#>    64   31 0.340082972 0.9565561379 0.344722980 0.8859681122   0.40714419
#>    31   95 0.344722980 0.8859681122 0.375150567 0.6908510357   0.39967149
#>    95   60 0.375150567 0.6908510357 0.624188854 0.7796834205   0.49460234
#>    57   31 0.166177808 0.7320144601 0.344722980 0.8859681122   0.24504549
#>    31   81 0.344722980 0.8859681122 0.215156565 0.9441810056   0.24908998
#>    75   73 0.307689589 0.6171812476 0.245701888 0.6279136322   0.26260483
#>    73   94 0.245701888 0.6279136322 0.147064800 0.5328545592   0.22879570
#>    94   47 0.147064800 0.5328545592 0.129878271 0.4581370937   0.17239662
#>    82   56 0.185703254 0.4260788618 0.240675041 0.4487273022   0.21419498
#>    32   66 0.991126601 0.2234403178 1.000000000 0.7079352223   2.54948857
#>    65   53 0.647178053 0.4068854098 0.431394748 0.3592419102   0.53354763
#>    79   37 0.457300490 0.2807282227 0.582485175 0.2349416669   0.54522344
#>    90   49 0.862394409 0.9764092843 0.851599074 0.9452961664   0.86782579
#>    49   72 0.851599074 0.9452961664 0.884158428 0.8450487322   0.90622619
#>    57   34 0.166177808 0.7320144601 0.274154468 0.6790806504   0.26528588
#>    34   95 0.274154468 0.6790806504 0.375150567 0.6908510357   0.31344189
#>    68   64 0.607236196 0.9833642085 0.340082972 0.9565561379   0.45799280
#>    57   42 0.166177808 0.7320144601 0.161218845 0.6971364773   0.22062411
#>    42   55 0.161218845 0.6971364773 0.204117058 0.6290607651   0.21716347
#>    55   73 0.204117058 0.6290607651 0.245701888 0.6279136322   0.22611801
#>    75   95 0.307689589 0.6171812476 0.375150567 0.6908510357   0.32668718
#>    47   52 0.129878271 0.4581370937 0.103672864 0.4754224578   0.11285855
#>    52   93 0.103672864 0.4754224578 0.073320062 0.4612126200   0.08430502
#>    50   78 0.679391963 0.6363686882 0.621446701 0.6576514968   0.63787174
#>    78   84 0.621446701 0.6576514968 0.598358301 0.6609461066   0.60504858
#>    84   46 0.598358301 0.6609461066 0.549339131 0.5796369553   0.56313765
#>    46   83 0.549339131 0.5796369553 0.674323558 0.4811310187   0.59095201
#>    53   46 0.431394748 0.3592419102 0.549339131 0.5796369553   0.47171758
#>    84   70 0.598358301 0.6609461066 0.384039832 0.6660281713   0.49145721
#>    70   75 0.384039832 0.6660281713 0.307689589 0.6171812476   0.36461059
#>    61   94 0.002896245 0.7026351919 0.147064800 0.5328545592   0.07933307
#>    94   55 0.147064800 0.5328545592 0.204117058 0.6290607651   0.11208297
#>    93   98 0.073320062 0.4612126200 0.041750174 0.4872383044   0.04764841
#>    98   45 0.041750174 0.4872383044 0.003881015 0.4696155084   0.03037367
#>    68   90 0.607236196 0.9833642085 0.862394409 0.9764092843   2.72050755
#>    98   61 0.041750174 0.4872383044 0.002896245 0.7026351919  -0.02724272
#>    61   59 0.002896245 0.7026351919 0.000000000 0.0797600839 -21.95902465
#>    59   51 0.000000000 0.0797600839 0.008634808 0.2419158597  -2.62476167
#>    94   52 0.147064800 0.5328545592 0.103672864 0.4754224578   0.13607985
#>    89   88 0.073733865 0.0797686021 0.057750178 0.0000000000   0.10134534
#>    62   89 0.046665551 0.1264752971 0.073733865 0.0797686021   0.15336783
#>    90   67 0.862394409 0.9764092843 0.783802145 0.9267604793   0.80608394
#>    67   96 0.783802145 0.9267604793 0.830077798 0.8920302043   0.82016454
#>    60   78 0.624188854 0.7796834205 0.621446701 0.6576514968   0.67629710
#>    59   62 0.000000000 0.0797600839 0.046665551 0.1264752971  -0.03663180
#>    94   98 0.147064800 0.5328545592 0.041750174 0.4872383044   0.08315025
#>    99   67 0.668227860 0.8731021291 0.783802145 0.9267604793   0.73048139
#>    95   84 0.375150567 0.6908510357 0.598358301 0.6609461066   0.49606316
#>     1    9 0.051546218 0.3815078300 0.136210111 0.3325412972   0.06199801
#>     1   21 0.051546218 0.3815078300 0.073139709 0.4182701574   0.10438641
#>     1   43 0.051546218 0.3815078300 0.041886486 0.4371951477   0.04634345
#>     1   45 0.051546218 0.3815078300 0.003881015 0.4696155084  -0.02536169
#>     1   51 0.051546218 0.3815078300 0.008634808 0.2419158597  -0.10561965
#>     2   74 0.617395031 0.1783063886 0.774725699 0.1178986432   0.68866553
#>     2   13 0.617395031 0.1783063886 0.813898244 0.1983189485   0.71477716
#>     2   29 0.617395031 0.1783063886 0.637692355 0.3635530003   0.70813191
#>     2   37 0.617395031 0.1783063886 0.582485175 0.2349416669   0.69268561
#>     2   86 0.617395031 0.1783063886 0.505427515 0.2218324095   0.55262078
#>     2   92 0.617395031 0.1783063886 0.442930085 0.0897716329   0.53332228
#>     2    4 0.617395031 0.1783063886 0.456812984 0.0634587788   0.53619154
#>     2    3 0.617395031 0.1783063886 0.552175139 0.0239003649   0.53673416
#>     2   87 0.617395031 0.1783063886 0.671945315 0.0437160870   0.60213916
#>     3   88 0.552175139 0.0239003649 0.057750178 0.0000000000   0.39562109
#>     3   71 0.552175139 0.0239003649 0.795106635 0.0004359111   0.42792527
#>     3   40 0.552175139 0.0239003649 0.681298621 0.0378172360   0.64769894
#>     3   10 0.552175139 0.0239003649 0.355200004 0.0473170282   0.43462324
#>     3   76 0.552175139 0.0239003649 0.319470241 0.0288505019   0.43216858
#>     5  100 0.693003274 0.5370731953 0.891743805 0.5325122165   0.79155383
#>     5   50 0.693003274 0.5370731953 0.679391963 0.6363686882   0.79390409
#>     5   36 0.693003274 0.5370731953 0.595882574 0.5901742426   0.65471158
#>     5   83 0.693003274 0.5370731953 0.674323558 0.4811310187   0.62529010
#>     5   22 0.693003274 0.5370731953 0.719541862 0.4140303035   0.75358163
#>     5   25 0.693003274 0.5370731953 0.884801800 0.4494671457   0.78895507
#>     6   44 0.213638605 0.9505520699 0.095749352 1.0000000000   0.26351058
#>     6   26 0.213638605 0.9505520699 0.170511389 0.9137596981   0.15488999
#>     6   81 0.213638605 0.9505520699 0.215156565 0.9441810056   0.18506149
#>     6   39 0.213638605 0.9505520699 0.333064705 0.9350749527   0.27583885
#>     6   64 0.213638605 0.9505520699 0.340082972 0.9565561379   0.27629306
#>     7   86 0.366305481 0.2274418018 0.505427515 0.2218324095   0.43386040
#>     7   79 0.366305481 0.2274418018 0.457300490 0.2807282227   0.43542724
#>     7   16 0.366305481 0.2274418018 0.422266874 0.3600743151   0.38669491
#>     7   63 0.366305481 0.2274418018 0.233757248 0.3776044701   0.32379579
#>     7   15 0.366305481 0.2274418018 0.314740855 0.1659393325   0.25826560
#>     7   91 0.366305481 0.2274418018 0.353479724 0.0914889201   0.38809614
#>     7   92 0.366305481 0.2274418018 0.442930085 0.0897716329   0.39945476
#>     8   80 0.843675470 0.1302412179 0.987003000 0.0360723231   0.88700940
#>     8   69 0.843675470 0.1302412179 0.929489754 0.1481877889   0.90241877
#>     8   13 0.843675470 0.1302412179 0.813898244 0.1983189485   0.87693614
#>     8   74 0.843675470 0.1302412179 0.774725699 0.1178986432   0.80394751
#>     8   71 0.843675470 0.1302412179 0.795106635 0.0004359111   0.81973716
#>     9   62 0.136210111 0.3325412972 0.046665551 0.1264752971   0.12324287
#>     9   15 0.136210111 0.3325412972 0.314740855 0.1659393325   0.17370276
#>     9   63 0.136210111 0.3325412972 0.233757248 0.3776044701   0.23109335
#>     9   82 0.136210111 0.3325412972 0.185703254 0.4260788618   0.17793780
#>    10   91 0.355200004 0.0473170282 0.353479724 0.0914889201   0.39782989
#>    11   68 0.563760744 0.8824759226 0.607236196 0.9833642085   0.61962175
#>    11   97 0.563760744 0.8824759226 0.542944764 0.9327711795   0.59969373
#>    11   33 0.563760744 0.8824759226 0.491173337 0.9222670774   0.52359461
#>    11   24 0.563760744 0.8824759226 0.464659455 0.8791219055   0.51429332
#>    11   60 0.563760744 0.8824759226 0.624188854 0.7796834205   0.51741602
#>    11   99 0.563760744 0.8824759226 0.668227860 0.8731021291   0.61279588
#>    12   38 0.931280129 0.5591655870 0.899256631 0.4782906157   0.93507612
#>    12   28 0.931280129 0.5591655870 0.976833765 0.4063398645   0.98376212
#>    12   66 0.931280129 0.5591655870 1.000000000 0.7079352223   1.15951586
#>    12   18 0.931280129 0.5591655870 0.966598831 0.6861169685   1.05031117
#>    12  100 0.931280129 0.5591655870 0.891743805 0.5325122165   0.83914313
#>    13   77 0.813898244 0.1983189485 0.943741881 0.2038872392   0.87926568
#>    13   54 0.813898244 0.1983189485 0.849510993 0.3516637626   0.87609349
#>    13   22 0.813898244 0.1983189485 0.719541862 0.4140303035   0.74269100
#>    14   44 0.112224478 0.8997614694 0.095749352 1.0000000000   0.05085242
#>    14   17 0.112224478 0.8997614694 0.103577689 0.8501654308  -0.11518667
#>    14   26 0.112224478 0.8997614694 0.170511389 0.9137596981   0.15080059
#>    14   85 0.112224478 0.8997614694 0.118735128 0.9116301637   0.14559048
#>    15   48 0.314740855 0.1659393325 0.154210921 0.0461049700   0.17486295
#>    15   76 0.314740855 0.1659393325 0.319470241 0.0288505019   0.24282890
#>    16   53 0.422266874 0.3600743151 0.431394748 0.3592419102   0.42255721
#>    16   41 0.422266874 0.3600743151 0.418953220 0.5946782097   0.43758784
#>    16   20 0.422266874 0.3600743151 0.304105854 0.4415576258   0.41593648
#>    17   61 0.103577689 0.8501654308 0.002896245 0.7026351919  -0.14583508
#>    17   57 0.103577689 0.8501654308 0.166177808 0.7320144601   0.07680107
#>    18   72 0.966598831 0.6861169685 0.884158428 0.8450487322   0.93519184
#>    18   23 0.966598831 0.6861169685 0.848202332 0.8184849075   0.91793041
#>    18   50 0.966598831 0.6861169685 0.679391963 0.6363686882   0.82066146
#>    19   54 0.902874551 0.4007002668 0.849510993 0.3516637626   0.84715429
#>    19   32 0.902874551 0.4007002668 0.991126601 0.2234403178   0.93883956
#>    19   28 0.902874551 0.4007002668 0.976833765 0.4063398645   0.94683379
#>    19   25 0.902874551 0.4007002668 0.884801800 0.4494671457   0.93699043
#>    20   75 0.304105854 0.4415576258 0.307689589 0.6171812476   0.34766473
#>    20   27 0.304105854 0.4415576258 0.218677603 0.4644675100   0.28199966
#>    20   56 0.304105854 0.4415576258 0.240675041 0.4487273022   0.28194125
#>    21   47 0.073139709 0.4182701574 0.129878271 0.4581370937   0.13124537
#>    21   35 0.073139709 0.4182701574 0.095636249 0.4614550545   0.10955538
#>    21   93 0.073139709 0.4182701574 0.073320062 0.4612126200   0.08471325
#>    22   65 0.719541862 0.4140303035 0.647178053 0.4068854098   0.68077002
#>    23   96 0.848202332 0.8184849075 0.830077798 0.8920302043   0.84733395
#>    23   99 0.848202332 0.8184849075 0.668227860 0.8731021291   0.75477791
#>    23   60 0.848202332 0.8184849075 0.624188854 0.7796834205   0.73904603
#>    24   64 0.464659455 0.8791219055 0.340082972 0.9565561379   0.41554066
#>    24   31 0.464659455 0.8791219055 0.344722980 0.8859681122   0.40714419
#>    24   95 0.464659455 0.8791219055 0.375150567 0.6908510357   0.39967149
#>    25   38 0.884801800 0.4494671457 0.899256631 0.4782906157   0.93708732
#>    26   31 0.170511389 0.9137596981 0.344722980 0.8859681122   0.24504549
#>    27   73 0.218677603 0.4644675100 0.245701888 0.6279136322   0.26260483
#>    27   94 0.218677603 0.4644675100 0.147064800 0.5328545592   0.22879570
#>    27   47 0.218677603 0.4644675100 0.129878271 0.4581370937   0.17239662
#>    27   82 0.218677603 0.4644675100 0.185703254 0.4260788618   0.17367448
#>    29   53 0.637692355 0.3635530003 0.431394748 0.3592419102   0.53354763
#>    29   79 0.637692355 0.3635530003 0.457300490 0.2807282227   0.53478538
#>    30   49 0.883280342 0.9698892753 0.851599074 0.9452961664   0.86782579
#>    30   72 0.883280342 0.9698892753 0.884158428 0.8450487322   0.90622619
#>    30   66 0.883280342 0.9698892753 1.000000000 0.7079352223   1.09889892
#>    31   39 0.344722980 0.8859681122 0.333064705 0.9350749527   0.40011456
#>    31   34 0.344722980 0.8859681122 0.274154468 0.6790806504   0.26528588
#>    32   77 0.991126601 0.2234403178 0.943741881 0.2038872392   0.93173963
#>    32   69 0.991126601 0.2234403178 0.929489754 0.1481877889   0.98843155
#>    32   80 0.991126601 0.2234403178 0.987003000 0.0360723231   1.02984490
#>    33   68 0.491173337 0.9222670774 0.607236196 0.9833642085   0.48854773
#>    34   42 0.274154468 0.6790806504 0.161218845 0.6971364773   0.22062411
#>    34   55 0.274154468 0.6790806504 0.204117058 0.6290607651   0.21716347
#>    34   73 0.274154468 0.6790806504 0.245701888 0.6279136322   0.22611801
#>    34   75 0.274154468 0.6790806504 0.307689589 0.6171812476   0.28011113
#>    35   52 0.095636249 0.4614550545 0.103672864 0.4754224578   0.11285855
#>    36   78 0.595882574 0.5901742426 0.621446701 0.6576514968   0.63787174
#>    36   84 0.595882574 0.5901742426 0.598358301 0.6609461066   0.60504858
#>    36   46 0.595882574 0.5901742426 0.549339131 0.5796369553   0.56313765
#>    40   74 0.681298621 0.0378172360 0.774725699 0.1178986432   0.74932606
#>    41   46 0.418953220 0.5946782097 0.549339131 0.5796369553   0.47171758
#>    41   84 0.418953220 0.5946782097 0.598358301 0.6609461066   0.49355285
#>    41   70 0.418953220 0.5946782097 0.384039832 0.6660281713   0.49145721
#>    42   61 0.161218845 0.6971364773 0.002896245 0.7026351919   0.08296639
#>    42   94 0.161218845 0.6971364773 0.147064800 0.5328545592   0.07933307
#>    43   98 0.041886486 0.4371951477 0.041750174 0.4872383044   0.04764841
#>    44   68 0.095749352 1.0000000000 0.607236196 0.9833642085   0.38086328
#>    44   90 0.095749352 1.0000000000 0.862394409 0.9764092843   2.72050755
#>    45   61 0.003881015 0.4696155084 0.002896245 0.7026351919  -0.02724272
#>    45   59 0.003881015 0.4696155084 0.000000000 0.0797600839 -21.95902465
#>    46   65 0.549339131 0.5796369553 0.647178053 0.4068854098   0.52421456
#>    48   88 0.154210921 0.0461049700 0.057750178 0.0000000000   0.10134534
#>    48   89 0.154210921 0.0461049700 0.073733865 0.0797686021   0.15336783
#>    49   67 0.851599074 0.9452961664 0.783802145 0.9267604793   0.80608394
#>    49   96 0.851599074 0.9452961664 0.830077798 0.8920302043   0.82016454
#>    52   98 0.103672864 0.4754224578 0.041750174 0.4872383044   0.08315025
#>    58   67 0.733120411 0.9548455915 0.783802145 0.9267604793   0.73048139
#>    58   90 0.733120411 0.9548455915 0.862394409 0.9764092843   0.79184830
#>    58   68 0.733120411 0.9548455915 0.607236196 0.9833642085   0.74332180
#>    58   99 0.733120411 0.9548455915 0.668227860 0.8731021291   0.66424091
#>    59   89 0.000000000 0.0797600839 0.073733865 0.0797686021   0.03687087
#>    60   84 0.624188854 0.7796834205 0.598358301 0.6609461066   0.49606316
#>    70   95 0.384039832 0.6660281713 0.375150567 0.6908510357   0.49251249
#>    30   90 0.883280342 0.9698892753 0.862394409 0.9764092843   0.86782579
#>    59   88 0.000000000 0.0797600839 0.057750178 0.0000000000   0.03687087
#>          my1          mx2         my2 bp1 bp2
#>   0.30190329   0.12324287  0.21568764   0   0
#>   0.37519348   0.13130671  0.39499864   0   0
#>   0.40928680   0.06480626  0.43977677   0   0
#>   0.39684856   0.03037367  0.46218555   0   0
#>   0.35342996  -2.62476167  0.30083665   0   0
#>   0.12884283   0.71188226  0.09667591   0   0
#>   0.19685002   0.80394751  0.15341535   0   0
#>   0.26209971   0.73746075  0.29337603   0   0
#>   0.26379215   0.54522344  0.32709116   0   0
#>   0.17745651   0.53289764  0.29339130   0   0
#>   0.12781252   0.43386040  0.17488260   0   0
#>   0.12215842   0.40628737  0.05361980   0   0
#>   0.12139971   0.43462324 -0.12475569   0   0
#>   0.09377315   0.62456533 -0.04177512   0   0
#>  -1.86349353   0.21214382 -0.19906214   0   0
#>  -2.53176762   0.42872703 -3.88796331   0   1
#>  -0.25641329   0.74932606  0.05299205   0   0
#>  -0.04177512   0.71188226  0.09667591   0   0
#>  -0.12475569   0.40628737  0.05361980   0   0
#>  -0.14540374   0.32180305  0.06813582   0   0
#>   0.05361980   0.39782989  0.07109670   0   0
#>   0.49907473   0.82790702  0.49603585   0   0
#>   0.60148522   0.82303065  0.66103931   0   0
#>   0.58240487   0.63787174  0.61284760   0   0
#>   0.52859361   0.59095201  0.50389227   0   0
#>   0.48575564   0.68077002  0.43668884   0   0
#>   0.49338518   0.80396946  0.42336546   0   0
#>   1.23470666   0.38086328  1.89471270   0   0
#>   0.97574334   0.14247284  0.96497887   0   0
#>   0.94037696   0.24908998  0.84641113   0   0
#>   0.96200545   0.27061078  0.89431070   0   0
#>   0.96551029   0.40011456  0.92505572   0   0
#>   0.21374285   0.53289764  0.29339130   0   0
#>   0.29696103   0.42255721  0.31279528   0   0
#>   0.32349994   0.32596649  0.34684249   0   0
#>   0.26565659   0.23109335  0.25526042   0   0
#>   0.15680465   0.27085275  0.09579926   0   0
#>   0.15573308   0.39782989  0.07109670   0   0
#>   0.04003802   1.82249148 -4.99738296   0   1
#>   0.06349148   1.02984490  0.12885883   0   0
#>   0.18534060   0.87926568  0.19071201   0   0
#>   0.06520906   0.74932606  0.05299205   0   0
#>   0.21568764  -0.03663180  0.16301852   0   0
#>   0.19376064   0.17486295  0.18587955   0   0
#>   0.37032493   0.22258873  0.41458863   0   0
#>   0.39499864   0.13124537  0.39588285   0   0
#>   0.06813582   0.27085275  0.09579926   0   0
#>   0.91821544   0.66424091  0.94289660   0   0
#>   0.92680295   0.48854773  1.06804247   0   0
#>   0.89530738   0.48854773  1.06804247   0   0
#>   0.87833988   0.41554066  0.93902608   0   0
#>   0.78607345   0.49460234  0.74947341   0   0
#>   0.84214393   0.73904603  0.78262768   0   0
#>   0.51088497   0.82790702  0.49603585   0   0
#>   0.49160711   0.93708732  0.44128246   0   0
#>   0.54399504   2.54948857  0.43722803   0   0
#>   0.59443901   0.93519184  0.77067312   0   0
#>   0.65318734   0.82303065  0.66103931   0   0
#>   0.19071201   0.98843155  0.16277914   0   0
#>   0.26468250   0.93173963  0.30016571   0   0
#>   0.29566385   0.80396946  0.42336546   0   0
#>   0.29337603   0.68681104  0.37550509   0   0
#>   0.94114759   0.14247284  0.96497887   0   0
#>   0.91385754  -0.14583508  0.91225626   0   0
#>   0.86748416   0.19374816  0.82228135   0   0
#>   0.88917841   0.14247284  0.96497887   0   0
#>   0.18587955   0.15336783  0.15711644   0   0
#>   0.09483247   0.21214382 -0.19906214   0   0
#>   0.31279528   0.53478538  0.34982518   0   0
#>   0.47761606   0.47171758  0.47941964   0   0
#>   0.47731025   0.34766473  0.52851715   0   0
#>   0.34684249   0.26843289  0.41012958   0   0
#>   0.91225626  -3.20995029  1.86903598   0   1
#>   0.76031908   0.08296639  0.72605396   0   0
#>   0.82228135   0.24504549  0.82105820   0   0
#>   0.77067312   1.09889892  0.90898250   0   0
#>   0.76171934   0.84733395  0.85727686   0   0
#>   0.67471713   0.74570215  0.74419973   0   0
#>   0.40778290   0.80396946  0.42336546   0   0
#>   0.30800718   0.93173963  0.30016571   0   0
#>   0.31198726   2.54948857  0.43722803   0   0
#>   0.44107571   0.93708732  0.44128246   0   0
#>   0.52851715   0.36461059  0.61230394   0   0
#>   0.52985710   0.26260483  0.54116172   0   0
#>   0.52963929   0.21419498  0.43496174   0   0
#>   0.41012958   0.22258873  0.41458863   0   0
#>   0.39588285   0.17367448  0.46976727   0   0
#>   0.42675202   0.11285855  0.46084140   0   0
#>   0.43969316   0.08430502  0.47727062   0   0
#>   0.43977677   0.04764841  0.46223261   0   0
#>   0.43668884   0.56992656  0.47721516   0   0
#>   0.37550509   0.53354763  0.40905528   0   0
#>   0.85727686   0.88511956  0.90077209   0   0
#>   0.83446731   0.75422892  0.83916157   0   0
#>   0.74419973   0.67629710  0.71746574   0   0
#>   0.93902608   0.45799280  1.12608600   0   0
#>   0.92551780   0.40011456  0.92505572   0   0
#>   0.79460602   0.31344189  0.78115892   0   0
#>   0.74947341   0.49606316  0.74537806   0   0
#>   0.82105820   0.26528588  0.79758475   0   0
#>   0.84641113   0.27061078  0.89431070   0   0
#>   0.54116172   0.28011113  0.64227395   0   0
#>   0.54675175   0.22282565  0.55294650   0   0
#>   0.48769237   0.13607985  0.49604596   0   0
#>   0.43496174   0.22258873  0.41458863   0   0
#>   0.43722803   4.61280167  0.39943898   0   1
#>   0.40905528   0.52421456  0.45132588   0   0
#>   0.32709116   0.53289764  0.29339130   0   0
#>   0.95709537   0.80608394  0.97851797   0   0
#>   0.90762731   0.88511956  0.90077209   0   0
#>   0.79758475   0.22062411  0.70648174   0   0
#>   0.78115892   0.32668718  0.66750737   0   0
#>   1.12608600   0.38086328  1.89471270   0   0
#>   0.70648174   0.08296639  0.72605396   0   0
#>   0.68483612   0.11208297  0.61861918   0   0
#>   0.67229805   0.22282565  0.55294650   0   0
#>   0.66750737   0.33297888  0.66174593   0   0
#>   0.46084140   0.13607985  0.49604596   0   0
#>   0.47727062   0.07551826  0.49603952   0   0
#>   0.61284760   0.67629710  0.71746574   0   0
#>   0.62528283   0.61838834  0.71876699   0   0
#>   0.62674895   0.49355285  0.66869982   0   0
#>   0.50389227   0.56992656  0.47721516   0   0
#>   0.47941964   0.52421456  0.45132588   0   0
#>   0.67437329   0.49251249  0.71887618   0   0
#>   0.61230394   0.33297888  0.66174593   0   0
#>   0.62144082   0.05507864  0.60084527   0   0
#>   0.61861918   0.22282565  0.55294650   0   0
#>   0.46223261   0.07551826  0.49603952   0   0
#>   0.46218555  -0.02724272  0.58599590   0   0
#>  73.82980870   0.74332180  1.29196808   0   0
#>   0.58599590   0.05507864  0.60084527   0   0
#>   0.49330946 -23.56444822  0.50077436   0   1
#>   0.30083665  -0.03663180  0.16301852   0   0
#>   0.49604596   0.08315025  0.53603610   0   0
#>   0.03275026   0.03687087  0.04566938   0   0
#>   0.15711644   0.03686580  0.08959906   0   0
#>   0.97851797   0.79184830  1.00105247   0   0
#>   0.92701615   0.75422892  0.83916157   0   0
#>   0.71746574   0.61838834  0.71876699   0   0
#>   0.16301852   0.03686580  0.08959906   0   0
#>   0.53603610   0.05507864  0.60084527   0   0
#>   0.89031119   0.75422892  0.83916157   0   0
#>   0.74537806   0.49251249  0.71887618   0   0
#>   0.30190329   0.10438641  0.37519348   0   0
#>   0.37519348   0.04634345  0.40928680   0   0
#>   0.40928680  -0.02536169  0.39684856   0   0
#>   0.39684856  -0.10561965  0.35342996   0   0
#>   0.35342996   0.06199801  0.30190329   0   0
#>   0.12884283   0.71477716  0.19685002   0   0
#>   0.19685002   0.70813191  0.26209971   0   0
#>   0.26209971   0.69268561  0.26379215   0   0
#>   0.26379215   0.55262078  0.17745651   0   0
#>   0.17745651   0.53332228  0.12781252   0   0
#>   0.12781252   0.53619154  0.12215842   0   0
#>   0.12215842   0.53673416  0.12139971   0   0
#>   0.12139971   0.60213916  0.09377315   0   0
#>   0.09377315   0.68866553  0.12884283   0   0
#>  -1.86349353   0.42792527 -2.53176762   0   0
#>  -2.53176762   0.64769894 -0.25641329   0   0
#>  -0.25641329   0.62456533 -0.04177512   0   0
#>  -0.12475569   0.43216858 -0.14540374   0   0
#>  -0.14540374   0.39562109 -1.86349353   0   0
#>   0.49907473   0.79390409  0.60148522   0   0
#>   0.60148522   0.65471158  0.58240487   0   0
#>   0.58240487   0.62529010  0.52859361   0   0
#>   0.52859361   0.75358163  0.48575564   0   0
#>   0.48575564   0.78895507  0.49338518   0   0
#>   0.49338518   0.79155383  0.49907473   0   0
#>   1.23470666   0.15488999  0.97574334   0   0
#>   0.97574334   0.18506149  0.94037696   0   0
#>   0.94037696   0.27583885  0.96200545   0   0
#>   0.96200545   0.27629306  0.96551029   0   0
#>   0.96551029   0.26351058  1.23470666   0   0
#>   0.17488260   0.43542724  0.21374285   0   0
#>   0.21374285   0.38669491  0.29696103   0   0
#>   0.29696103   0.32379579  0.32349994   0   0
#>   0.32349994   0.25826560  0.26565659   0   0
#>   0.26565659   0.38809614  0.15680465   0   0
#>   0.15680465   0.39945476  0.15573308   0   0
#>   0.15573308   0.43386040  0.17488260   0   0
#>   0.04003802   0.90241877  0.06349148   0   0
#>   0.06349148   0.87693614  0.18534060   0   0
#>   0.18534060   0.80394751  0.15341535   0   0
#>   0.15341535   0.81973716  0.06520906   0   0
#>   0.06520906   0.88700940  0.04003802   0   0
#>   0.21568764   0.17370276  0.19376064   0   0
#>   0.19376064   0.23109335  0.25526042   0   0
#>   0.25526042   0.17793780  0.37032493   0   0
#>   0.37032493   0.13130671  0.39499864   0   0
#>   0.07109670   0.32180305  0.06813582   0   0
#>   0.91821544   0.59969373  0.92680295   0   0
#>   0.92680295   0.52359461  0.89530738   0   0
#>   0.89530738   0.51429332  0.87833988   0   0
#>   0.87833988   0.51741602  0.78607345   0   0
#>   0.78607345   0.61279588  0.84214393   0   0
#>   0.84214393   0.61962175  0.91821544   0   0
#>   0.51088497   0.98376212  0.49160711   0   0
#>   0.49160711   1.15951586  0.54399504   0   0
#>   0.54399504   1.05031117  0.59443901   0   0
#>   0.59443901   0.83914313  0.65318734   0   0
#>   0.65318734   0.93507612  0.51088497   0   0
#>   0.19071201   0.87609349  0.26468250   0   0
#>   0.26468250   0.74269100  0.29566385   0   0
#>   0.29566385   0.73746075  0.29337603   0   0
#>   0.94114759  -0.11518667  0.91385754   0   0
#>   0.91385754   0.15080059  0.86748416   0   0
#>   0.86748416   0.14559048  0.88917841   0   0
#>   0.88917841   0.05085242  0.94114759   0   0
#>   0.18587955   0.24282890  0.09483247   0   0
#>   0.09483247   0.27085275  0.09579926   0   0
#>   0.31279528   0.43758784  0.47761606   0   0
#>   0.47761606   0.41593648  0.47731025   0   0
#>   0.47731025   0.32596649  0.34684249   0   0
#>   0.91225626   0.07680107  0.76031908   0   0
#>   0.76031908   0.19374816  0.82228135   0   0
#>   0.77067312   0.91793041  0.76171934   0   0
#>   0.76171934   0.82066146  0.67471713   0   0
#>   0.67471713   0.82303065  0.66103931   0   0
#>   0.40778290   0.93883956  0.30800718   0   0
#>   0.30800718   0.94683379  0.31198726   0   0
#>   0.31198726   0.93699043  0.44107571   0   0
#>   0.44107571   0.84715429  0.40778290   0   0
#>   0.52851715   0.28199966  0.52985710   0   0
#>   0.52985710   0.28194125  0.52963929   0   0
#>   0.52963929   0.26843289  0.41012958   0   0
#>   0.39588285   0.10955538  0.42675202   0   0
#>   0.42675202   0.08471325  0.43969316   0   0
#>   0.43969316   0.06480626  0.43977677   0   0
#>   0.43668884   0.68681104  0.37550509   0   0
#>   0.85727686   0.75477791  0.83446731   0   0
#>   0.83446731   0.73904603  0.78262768   0   0
#>   0.78262768   0.74570215  0.74419973   0   0
#>   0.93902608   0.40714419  0.92551780   0   0
#>   0.92551780   0.39967149  0.79460602   0   0
#>   0.79460602   0.49460234  0.74947341   0   0
#>   0.44128246   0.82790702  0.49603585   0   0
#>   0.82105820   0.24908998  0.84641113   0   0
#>   0.54116172   0.22879570  0.54675175   0   0
#>   0.54675175   0.17239662  0.48769237   0   0
#>   0.48769237   0.17367448  0.46976727   0   0
#>   0.46976727   0.21419498  0.43496174   0   0
#>   0.40905528   0.53478538  0.34982518   0   0
#>   0.34982518   0.54522344  0.32709116   0   0
#>   0.95709537   0.90622619  0.90762731   0   0
#>   0.90762731   1.09889892  0.90898250   0   0
#>   0.90898250   4.28400604  2.32818030   0   1
#>   0.92505572   0.27061078  0.89431070   0   0
#>   0.79758475   0.31344189  0.78115892   0   0
#>   0.30016571   0.98843155  0.16277914   0   0
#>   0.16277914   1.02984490  0.12885883   0   0
#>   0.12885883   6.36435191  0.01145684   0   1
#>   1.06804247   0.45799280  1.12608600   0   0
#>   0.70648174   0.21716347  0.68483612   0   0
#>   0.68483612   0.22611801  0.67229805   0   0
#>   0.67229805   0.28011113  0.64227395   0   0
#>   0.64227395   0.32668718  0.66750737   0   0
#>   0.46084140   0.08430502  0.47727062   0   0
#>   0.61284760   0.60504858  0.62528283   0   0
#>   0.62528283   0.56313765  0.62674895   0   0
#>   0.62674895   0.59095201  0.50389227   0   0
#>   0.05299205   0.71188226  0.09667591   0   0
#>   0.47941964   0.49355285  0.66869982   0   0
#>   0.66869982   0.49145721  0.67437329   0   0
#>   0.67437329   0.36461059  0.61230394   0   0
#>   0.72605396   0.07933307  0.62144082   0   0
#>   0.62144082   0.11208297  0.61861918   0   0
#>   0.46223261   0.03037367  0.46218555   0   0
#>   1.89471270   2.72050755 73.82980870   0   0
#>  73.82980870   2.76060727 75.13295937   0   1
#>   0.58599590 -21.95902465  0.49330946   0   0
#>   0.49330946  -2.62476167  0.30083665   0   0
#>   0.45132588   0.56992656  0.47721516   0   0
#>   0.03275026   0.21214382 -0.19906214   0   0
#>   0.15711644   0.10134534  0.03275026   0   0
#>   0.97851797   0.82016454  0.92701615   0   0
#>   0.92701615   0.88511956  0.90077209   0   0
#>   0.53603610   0.07551826  0.49603952   0   0
#>   0.89031119   0.79184830  1.00105247   0   0
#>   1.00105247   0.74332180  1.29196808   0   0
#>   1.29196808   0.66424091  0.94289660   0   0
#>   0.94289660   0.73048139  0.89031119   0   0
#>   0.04566938   0.03686580  0.08959906   0   0
#>   0.74537806   0.61838834  0.71876699   0   0
#>   0.71887618   0.33297888  0.66174593   0   0
#>   0.95709537  14.48713518 44.58464037   0   1
#>   0.04566938  -8.18857071 -5.90995018   0   1
#> 
#> $x
#>               [,1]         [,2]
#>   [1,] 0.051546218 0.3815078300
#>   [2,] 0.617395031 0.1783063886
#>   [3,] 0.552175139 0.0239003649
#>   [4,] 0.456812984 0.0634587788
#>   [5,] 0.693003274 0.5370731953
#>   [6,] 0.213638605 0.9505520699
#>   [7,] 0.366305481 0.2274418018
#>   [8,] 0.843675470 0.1302412179
#>   [9,] 0.136210111 0.3325412972
#>  [10,] 0.355200004 0.0473170282
#>  [11,] 0.563760744 0.8824759226
#>  [12,] 0.931280129 0.5591655870
#>  [13,] 0.813898244 0.1983189485
#>  [14,] 0.112224478 0.8997614694
#>  [15,] 0.314740855 0.1659393325
#>  [16,] 0.422266874 0.3600743151
#>  [17,] 0.103577689 0.8501654308
#>  [18,] 0.966598831 0.6861169685
#>  [19,] 0.902874551 0.4007002668
#>  [20,] 0.304105854 0.4415576258
#>  [21,] 0.073139709 0.4182701574
#>  [22,] 0.719541862 0.4140303035
#>  [23,] 0.848202332 0.8184849075
#>  [24,] 0.464659455 0.8791219055
#>  [25,] 0.884801800 0.4494671457
#>  [26,] 0.170511389 0.9137596981
#>  [27,] 0.218677603 0.4644675100
#>  [28,] 0.976833765 0.4063398645
#>  [29,] 0.637692355 0.3635530003
#>  [30,] 0.883280342 0.9698892753
#>  [31,] 0.344722980 0.8859681122
#>  [32,] 0.991126601 0.2234403178
#>  [33,] 0.491173337 0.9222670774
#>  [34,] 0.274154468 0.6790806504
#>  [35,] 0.095636249 0.4614550545
#>  [36,] 0.595882574 0.5901742426
#>  [37,] 0.582485175 0.2349416669
#>  [38,] 0.899256631 0.4782906157
#>  [39,] 0.333064705 0.9350749527
#>  [40,] 0.681298621 0.0378172360
#>  [41,] 0.418953220 0.5946782097
#>  [42,] 0.161218845 0.6971364773
#>  [43,] 0.041886486 0.4371951477
#>  [44,] 0.095749352 1.0000000000
#>  [45,] 0.003881015 0.4696155084
#>  [46,] 0.549339131 0.5796369553
#>  [47,] 0.129878271 0.4581370937
#>  [48,] 0.154210921 0.0461049700
#>  [49,] 0.851599074 0.9452961664
#>  [50,] 0.679391963 0.6363686882
#>  [51,] 0.008634808 0.2419158597
#>  [52,] 0.103672864 0.4754224578
#>  [53,] 0.431394748 0.3592419102
#>  [54,] 0.849510993 0.3516637626
#>  [55,] 0.204117058 0.6290607651
#>  [56,] 0.240675041 0.4487273022
#>  [57,] 0.166177808 0.7320144601
#>  [58,] 0.733120411 0.9548455915
#>  [59,] 0.000000000 0.0797600839
#>  [60,] 0.624188854 0.7796834205
#>  [61,] 0.002896245 0.7026351919
#>  [62,] 0.046665551 0.1264752971
#>  [63,] 0.233757248 0.3776044701
#>  [64,] 0.340082972 0.9565561379
#>  [65,] 0.647178053 0.4068854098
#>  [66,] 1.000000000 0.7079352223
#>  [67,] 0.783802145 0.9267604793
#>  [68,] 0.607236196 0.9833642085
#>  [69,] 0.929489754 0.1481877889
#>  [70,] 0.384039832 0.6660281713
#>  [71,] 0.795106635 0.0004359111
#>  [72,] 0.884158428 0.8450487322
#>  [73,] 0.245701888 0.6279136322
#>  [74,] 0.774725699 0.1178986432
#>  [75,] 0.307689589 0.6171812476
#>  [76,] 0.319470241 0.0288505019
#>  [77,] 0.943741881 0.2038872392
#>  [78,] 0.621446701 0.6576514968
#>  [79,] 0.457300490 0.2807282227
#>  [80,] 0.987003000 0.0360723231
#>  [81,] 0.215156565 0.9441810056
#>  [82,] 0.185703254 0.4260788618
#>  [83,] 0.674323558 0.4811310187
#>  [84,] 0.598358301 0.6609461066
#>  [85,] 0.118735128 0.9116301637
#>  [86,] 0.505427515 0.2218324095
#>  [87,] 0.671945315 0.0437160870
#>  [88,] 0.057750178 0.0000000000
#>  [89,] 0.073733865 0.0797686021
#>  [90,] 0.862394409 0.9764092843
#>  [91,] 0.353479724 0.0914889201
#>  [92,] 0.442930085 0.0897716329
#>  [93,] 0.073320062 0.4612126200
#>  [94,] 0.147064800 0.5328545592
#>  [95,] 0.375150567 0.6908510357
#>  [96,] 0.830077798 0.8920302043
#>  [97,] 0.542944764 0.9327711795
#>  [98,] 0.041750174 0.4872383044
#>  [99,] 0.668227860 0.8731021291
#> [100,] 0.891743805 0.5325122165
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.05154622,0.3815078) [5]: 9 21 43 45 51 
#> 2: (0.617395,0.1783064) [9]: 3 4 13 29 37 74 86 87 92 
#> 3: (0.5521751,0.02390036) [8]: 2 4 10 40 71 76 87 88 
#> 4: (0.456813,0.06345878) [4]: 2 3 10 92 
#> 5: (0.6930033,0.5370732) [6]: 22 25 36 50 83 100 
#> 6: (0.2136386,0.9505521) [5]: 26 39 44 64 81 
#> 7: (0.3663055,0.2274418) [7]: 15 16 63 79 86 91 92 
#> 8: (0.8436755,0.1302412) [5]: 13 69 71 74 80 
#> 9: (0.1362101,0.3325413) [7]: 1 15 21 51 62 63 82 
#> 10: (0.3552,0.04731703) [5]: 3 4 76 91 92 
#> 11: (0.5637607,0.8824759) [6]: 24 33 60 68 97 99 
#> 12: (0.9312801,0.5591656) [5]: 18 28 38 66 100 
#> 13: (0.8138982,0.1983189) [8]: 2 8 22 29 54 69 74 77 
#> 14: (0.1122245,0.8997615) [4]: 17 26 44 85 
#> 15: (0.3147409,0.1659393) [7]: 7 9 48 62 63 76 91 
#> 16: (0.4222669,0.3600743) [6]: 7 20 41 53 63 79 
#> 17: (0.1035777,0.8501654) [5]: 14 26 44 57 61 
#> 18: (0.9665988,0.686117) [6]: 12 23 50 66 72 100 
#> 19: (0.9028746,0.4007003) [4]: 25 28 32 54 
#> 20: (0.3041059,0.4415576) [6]: 16 27 41 56 63 75 
#> 21: (0.07313971,0.4182702) [7]: 1 9 35 43 47 82 93 
#> 22: (0.7195419,0.4140303) [7]: 5 13 25 29 54 65 83 
#> 23: (0.8482023,0.8184849) [6]: 18 50 60 72 96 99 
#> 24: (0.4646595,0.8791219) [6]: 11 31 33 60 64 95 
#> 25: (0.8848018,0.4494671) [7]: 5 19 22 28 38 54 100 
#> 26: (0.1705114,0.9137597) [8]: 6 14 17 31 44 57 81 85 
#> 27: (0.2186776,0.4644675) [7]: 20 47 56 73 75 82 94 
#> 28: (0.9768338,0.4063399) [6]: 12 19 25 32 38 66 
#> 29: (0.6376924,0.363553) [7]: 2 13 22 37 53 65 79 
#> 30: (0.8832803,0.9698893) [4]: 49 66 72 90 
#> 31: (0.344723,0.8859681) [8]: 24 26 34 39 57 64 81 95 
#> 32: (0.9911266,0.2234403) [7]: 19 28 54 66 69 77 80 
#> 33: (0.4911733,0.9222671) [5]: 11 24 64 68 97 
#> 34: (0.2741545,0.6790807) [7]: 31 42 55 57 73 75 95 
#> 35: (0.09563625,0.4614551) [4]: 21 47 52 93 
#> 36: (0.5958826,0.5901742) [6]: 5 46 50 78 83 84 
#> 37: (0.5824852,0.2349417) [4]: 2 29 79 86 
#> 38: (0.8992566,0.4782906) [4]: 12 25 28 100 
#> 39: (0.3330647,0.935075) [4]: 6 31 64 81 
#> 40: (0.6812986,0.03781724) [4]: 3 71 74 87 
#> 41: (0.4189532,0.5946782) [7]: 16 20 46 53 70 75 84 
#> 42: (0.1612188,0.6971365) [5]: 34 55 57 61 94 
#> 43: (0.04188649,0.4371951) [5]: 1 21 45 93 98 
#> 44: (0.09574935,1) [9]: 6 14 17 26 61 64 68 85 90 
#> 45: (0.003881015,0.4696155) [6]: 1 43 51 59 61 98 
#> 46: (0.5493391,0.579637) [6]: 36 41 53 65 83 84 
#> 47: (0.1298783,0.4581371) [6]: 21 27 35 52 82 94 
#> 48: (0.1542109,0.04610497) [5]: 15 62 76 88 89 
#> 49: (0.8515991,0.9452962) [5]: 30 67 72 90 96 
#> 50: (0.679392,0.6363687) [7]: 5 18 23 36 60 78 100 
#> 51: (0.008634808,0.2419159) [5]: 1 9 45 59 62 
#> 52: (0.1036729,0.4754225) [5]: 35 47 93 94 98 
#> 53: (0.4313947,0.3592419) [6]: 16 29 41 46 65 79 
#> 54: (0.849511,0.3516638) [6]: 13 19 22 25 32 77 
#> 55: (0.2041171,0.6290608) [4]: 34 42 73 94 
#> 56: (0.240675,0.4487273) [4]: 20 27 63 82 
#> 57: (0.1661778,0.7320145) [6]: 17 26 31 34 42 61 
#> 58: (0.7331204,0.9548456) [4]: 67 68 90 99 
#> 59: (0,0.07976008) [6]: 45 51 61 62 88 89 
#> 60: (0.6241889,0.7796834) [8]: 11 23 24 50 78 84 95 99 
#> 61: (0.002896245,0.7026352) [8]: 17 42 44 45 57 59 94 98 
#> 62: (0.04666555,0.1264753) [6]: 9 15 48 51 59 89 
#> 63: (0.2337572,0.3776045) [7]: 7 9 15 16 20 56 82 
#> 64: (0.340083,0.9565561) [7]: 6 24 31 33 39 44 68 
#> 65: (0.6471781,0.4068854) [5]: 22 29 46 53 83 
#> 66: (1,0.7079352) [6]: 12 18 28 30 32 72 
#> 67: (0.7838021,0.9267605) [5]: 49 58 90 96 99 
#> 68: (0.6072362,0.9833642) [8]: 11 33 44 58 64 90 97 99 
#> 69: (0.9294898,0.1481878) [5]: 8 13 32 77 80 
#> 70: (0.3840398,0.6660282) [4]: 41 75 84 95 
#> 71: (0.7951066,0.0004359111) [6]: 3 8 40 74 80 88 
#> 72: (0.8841584,0.8450487) [6]: 18 23 30 49 66 96 
#> 73: (0.2457019,0.6279136) [5]: 27 34 55 75 94 
#> 74: (0.7747257,0.1178986) [6]: 2 8 13 40 71 87 
#> 75: (0.3076896,0.6171812) [7]: 20 27 34 41 70 73 95 
#> 76: (0.3194702,0.0288505) [6]: 3 10 15 48 88 91 
#> 77: (0.9437419,0.2038872) [4]: 13 32 54 69 
#> 78: (0.6214467,0.6576515) [4]: 36 50 60 84 
#> 79: (0.4573005,0.2807282) [6]: 7 16 29 37 53 86 
#> 80: (0.987003,0.03607232) [4]: 8 32 69 71 
#> 81: (0.2151566,0.944181) [4]: 6 26 31 39 
#> 82: (0.1857033,0.4260789) [6]: 9 21 27 47 56 63 
#> 83: (0.6743236,0.481131) [5]: 5 22 36 46 65 
#> 84: (0.5983583,0.6609461) [7]: 36 41 46 60 70 78 95 
#> 85: (0.1187351,0.9116302) [3]: 14 26 44 
#> 86: (0.5054275,0.2218324) [5]: 2 7 37 79 92 
#> 87: (0.6719453,0.04371609) [4]: 2 3 40 74 
#> 88: (0.05775018,0) [6]: 3 48 59 71 76 89 
#> 89: (0.07373386,0.0797686) [4]: 48 59 62 88 
#> 90: (0.8623944,0.9764093) [6]: 30 44 49 58 67 68 
#> 91: (0.3534797,0.09148892) [5]: 7 10 15 76 92 
#> 92: (0.4429301,0.08977163) [6]: 2 4 7 10 86 91 
#> 93: (0.07332006,0.4612126) [5]: 21 35 43 52 98 
#> 94: (0.1470648,0.5328546) [8]: 27 42 47 52 55 61 73 98 
#> 95: (0.3751506,0.690851) [7]: 24 31 34 60 70 75 84 
#> 96: (0.8300778,0.8920302) [5]: 23 49 67 72 99 
#> 97: (0.5429448,0.9327712) [3]: 11 33 68 
#> 98: (0.04175017,0.4872383) [6]: 43 45 52 61 93 94 
#> 99: (0.6682279,0.8731021) [7]: 11 23 58 60 67 68 96 
#> 100: (0.8917438,0.5325122) [6]: 5 12 18 25 38 50 
#> number of nodes: 100 
#> number of arcs: 287 
#> number of boundary nodes: 10 
#> boundary nodes:  30 32 44 59 61 66 71 80 88 90 
#> number of triangles: 188 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.156487788 0.106429851 0.036536546 0.049954936 0.227749267 0.126755114 
#>                                                                         
#> 0.089453417 0.241559117 0.139959665 0.078164798 0.146102627 0.029750650 
#>                                                                         
#> 0.103241507 0.121398344 0.263305418 0.737356586 0.119789930 0.011058064 
#>                                                                         
#> 0.102887093 0.040219753 0.097462611 0.083334718 0.236388391 0.095434506 
#>                                                                         
#> 0.134325770 0.080914776 0.169016617 0.248165845 0.114134797 0.054024510 
#>                                                                         
#> 0.118259248 0.022598615 0.076058710 0.086736132 0.189322965 0.226628495 
#>                                                                         
#> 0.083925943 0.089466843 0.195177275 0.126006552 0.125994168 0.119217767 
#>                                                                         
#> 0.121543658 0.270964534 0.068256524 0.112834071 0.071275637 0.126006782 
#>                                                                         
#> 0.081811019 0.052826289 0.050640811 0.187983088 0.103278696 0.054739607 
#>                                                                         
#> 0.105807006 0.302483775 0.039895793 0.170873330 0.057493929 0.175263690 
#>                                                                         
#> 0.144158109 0.096162882 0.091310316 0.150038932 0.092327412 0.051820036 
#>                                                                         
#> 0.134258691 0.166157635 0.082677123 0.235764804 0.191404887 0.095073300 
#>                                                                         
#> 0.311524523 0.165903617 0.181796896 0.179497589 0.044704336 0.248320877 
#>                                                                         
#> 0.103975684 0.191039870 0.183457159 0.101635845 0.113516441 0.176761476 
#>                                                                         
#> 0.027048871 0.071458471 0.064375142 0.034402398 0.022317504 0.039558926 
#>                                                                         
#> 0.079052444 0.044358496 0.071637792 0.162952983 0.153578956 0.154932366 
#>                                                                         
#> 0.070740364 0.197475344 0.264407377 0.235754333 0.142042940 0.062909929 
#>                                                                         
#> 0.136987234 0.076668614 0.059454598 0.484576155 0.220980402 0.133295214 
#>                                                                         
#> 0.032932740 0.105402370 0.120253678 0.101679664 0.268494913 0.035228753 
#>                                                                         
#> 0.080464646 0.041600649 0.099891047 0.031392789 0.033514356 0.061730149 
#>                                                                         
#> 0.023322279 0.094942388 0.159136817 0.249969705 0.214378716 0.090638742 
#>                                                                         
#> 0.222733104 0.111850767 0.040914472 0.041768842 0.255252981 0.218873130 
#>                                                                         
#> 0.622881841 0.162385515 0.071981292 0.081354214 0.053983414 0.092961001 
#>                                                                         
#> 0.057858690 0.122062729 0.066030181 0.114769391 0.127423051 0.225202126 
#>                                                                         
#> 0.097804377 0.042635051 0.056518915 0.100174520 0.146038718 0.168529033 
#>                                                                         
#> 0.197519658 0.186355275 0.066530089 0.120130093 0.195643606 0.197424840 
#>                                                                         
#> 0.167615198 0.145224939 0.495002292 0.244062067 0.129871294 0.198362154 
#>                                                                         
#> 0.232757542 0.198792859 0.100224063 0.110689438 0.058978462 0.125872355 
#>                                                                         
#> 0.210858944 0.127839640 0.056688935 0.006549402 0.120424809 0.126586835 
#>                                                                         
#> 0.139235073 0.105449203 0.143955066 0.200294436 0.080258734 0.136556531 
#>                                                                         
#> 0.157557625 0.171495078 0.087670809 0.074305185 0.070045771 0.138594190 
#>                                                                         
#> 0.224680717 0.244191402 0.107452936 0.105824606 0.044205377 0.109857003 
#>                                                                         
#> 0.054432691 0.082778425 0.099158030 0.119238647 0.104886826 0.086984283 
#>                                                                         
#> 0.159470484 0.163874419 0.131772774 0.047681475 0.129962979 0.157425855 
#>                                                                         
#> 0.235445356 0.101583428 0.050344156 0.059944261 0.013537151 0.200325071 
#>                                                                         
#> 0.137170385 0.009165750 0.234627295 0.143532423 0.178611099 0.133710234 
#>                                                                         
#> 0.179041128 0.177592236 0.291483578 0.072472395 0.198013923 0.074173921 
#>                                                                         
#> 0.052008007 0.175660182 0.088446870 0.063834726 0.069344337 0.048693220 
#>                                                                         
#> 0.042942841 0.072715681 0.075745689 0.188079375 0.227349057 0.146681153 
#>                                                                         
#> 0.120131714 0.208465252 0.032244915 0.176414429 0.165665165 0.099021119 
#>                                                                         
#> 0.089024690 0.050606284 0.206342648 0.198497276 0.040106405 0.124843631 
#>                                                                         
#> 0.286781109 0.050471746 0.218591713 0.051260458 0.097273038 0.187413365 
#>                                                                         
#> 0.131161910 0.114369873 0.086065253 0.058545820 0.070399861 0.016114451 
#>                                                                         
#> 0.072157497 0.070815154 0.047721342 0.123051415 0.131250620 0.191252758 
#>                                                                         
#> 0.079434008 0.158418059 0.164890526 0.050043342 0.511757307 0.767007930 
#>                                                                         
#> 0.233021764 0.389874742 0.198533501 0.106912783 0.087234149 0.070285100 
#>                                                                         
#> 0.057449352 0.063039937 0.057943177 0.131060136 0.129074192 0.104369712 
#>                                                             
#> 0.073733865 0.121514473 0.026366525 0.021879961 0.098472098 
#> 
#> $alpha
#> [1] 0.3268804
#> 
#> attr(,"class")
#> [1] "scree"