Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1           y1          x2           y2          mx1
#>    73   53 0.176442448 0.4719123770 0.257069450 0.4998742742   0.20446911
#>    53   52 0.257069450 0.4998742742 0.194315905 0.4551471535   0.24120247
#>    32   86 0.339546529 0.5463479760 0.318888661 0.6340997645   0.26854045
#>    53   14 0.257069450 0.4998742742 0.287310817 0.4642783664   0.30548929
#>    90   75 0.065061541 0.5679794213 0.083136903 0.4255804275   0.10405024
#>    73    1 0.176442448 0.4719123770 0.181984080 0.5734872478   0.12564784
#>    90    1 0.065061541 0.5679794213 0.181984080 0.5734872478   0.12407361
#>    75   42 0.083136903 0.4255804275 0.177346563 0.3690873285   0.14395962
#>    42   52 0.177346563 0.3690873285 0.194315905 0.4551471535   0.14475221
#>    83   14 0.241985792 0.3553199246 0.287310817 0.4642783664   0.24490810
#>    86    6 0.318888661 0.6340997645 0.206508258 0.6998410128   0.23979767
#>    86    1 0.318888661 0.6340997645 0.181984080 0.5734872478   0.26128977
#>    83   69 0.241985792 0.3553199246 0.318899919 0.3235354913   0.30294091
#>    42   83 0.177346563 0.3690873285 0.241985792 0.3553199246   0.21890806
#>     4   11 0.066494593 0.4173131538 0.140260446 0.3026416731   0.10487340
#>    93    4 0.008134985 0.4799288952 0.066494593 0.4173131538   0.05372072
#>    49    9 0.061971719 0.6669235077 0.072727101 0.5894686361   0.13557497
#>    90   93 0.065061541 0.5679794213 0.008134985 0.4799288952   0.07788620
#>    90    9 0.065061541 0.5679794213 0.072727101 0.5894686361  -0.01700321
#>     6   49 0.206508258 0.6998410128 0.061971719 0.6669235077   0.14259674
#>    11   42 0.140260446 0.3026416731 0.177346563 0.3690873285   0.10978689
#>    94   32 0.442146235 0.4385407696 0.339546529 0.5463479760   0.36792320
#>    45   69 0.166294973 0.2430968603 0.318899919 0.3235354913   0.25081360
#>     6   19 0.206508258 0.6998410128 0.345653391 0.7770155566   0.25928351
#>    20   94 0.375491760 0.3422080701 0.442146235 0.4385407696   0.36017742
#>    69   20 0.318899919 0.3235354913 0.375491760 0.3422080701   0.32540571
#>    22   32 0.352466091 0.7119194799 0.339546529 0.5463479760   0.50492362
#>    58   50 0.393377769 0.2399969693 0.391651093 0.2772765735   0.33481053
#>    11   45 0.140260446 0.3026416731 0.166294973 0.2430968603   0.20782904
#>    37   88 0.000000000 0.2540724224 0.013989242 0.2089244778   0.07086250
#>    22   19 0.352466091 0.7119194799 0.345653391 0.7770155566   0.27691874
#>    50   99 0.391651093 0.2772765735 0.417725664 0.2638078999   0.42991107
#>    50   20 0.391651093 0.2772765735 0.375491760 0.3422080701   0.35700863
#>     6   60 0.206508258 0.6998410128 0.203687760 0.8353304547   0.11681466
#>    60   19 0.203687760 0.8353304547 0.345653391 0.7770155566   0.26182639
#>     2   44 0.345190939 0.8258541414 0.338935752 0.8078522612   0.38534800
#>    80   94 0.485607825 0.3309353779 0.442146235 0.4385407696   0.43425587
#>    58   99 0.393377769 0.2399969693 0.417725664 0.2638078999   0.39868634
#>    60   44 0.203687760 0.8353304547 0.338935752 0.8078522612   0.27507800
#>    55   59 0.464376280 0.8182976599 0.412139647 0.8591150901   0.40537424
#>    88   11 0.013989242 0.2089244778 0.140260446 0.3026416731   0.08081083
#>     8   93 0.018250055 0.2988399633 0.008134985 0.4799288952  -0.23401688
#>    50   69 0.391651093 0.2772765735 0.318899919 0.3235354913   0.32968854
#>    80   12 0.485607825 0.3309353779 0.570160514 0.4049472651   0.50030311
#>    45   98 0.166294973 0.2430968603 0.133115972 0.1867342914   0.19378182
#>    98   11 0.133115972 0.1867342914 0.140260446 0.3026416731   0.08289874
#>    23   49 0.028039999 0.6894806926 0.061971719 0.6669235077   0.10209650
#>    62   71 0.177772755 0.9035000303 0.053500904 0.8572904488   0.12851437
#>    80   99 0.485607825 0.3309353779 0.417725664 0.2638078999   0.45228025
#>    58   68 0.393377769 0.2399969693 0.325025648 0.1960531145   0.38889263
#>     2   10 0.345190939 0.8258541414 0.265439129 0.9374370246   0.27647806
#>    55   22 0.464376280 0.8182976599 0.352466091 0.7119194799   0.47040019
#>    80   27 0.485607825 0.3309353779 0.481699473 0.2591508101   0.54907019
#>    21   78 0.183239139 0.1291845698 0.200491547 0.1117229888   0.24790536
#>    68   45 0.325025648 0.1960531145 0.166294973 0.2430968603   0.23856371
#>    98   88 0.133115972 0.1867342914 0.013989242 0.2089244778   0.06651450
#>    59    2 0.412139647 0.8591150901 0.345190939 0.8258541414   0.39880818
#>    94   12 0.442146235 0.4385407696 0.570160514 0.4049472651   0.54796085
#>    10   62 0.265439129 0.9374370246 0.177772755 0.9035000303   0.23487320
#>    68   46 0.325025648 0.1960531145 0.285892575 0.0901326655   0.34124426
#>    66   78 0.147822819 0.0219815501 0.200491547 0.1117229888   0.15123184
#>    47   48 0.480817562 0.8315532035 0.460583336 0.8479285424   0.46566682
#>    10   70 0.265439129 0.9374370246 0.218576046 0.9523664380   0.22852761
#>    59   24 0.412139647 0.8591150901 0.352022916 0.9199484307   0.36411327
#>    24   10 0.352022916 0.9199484307 0.265439129 0.9374370246   0.29820124
#>    29   72 0.509456264 0.8118322704 0.503676928 0.8300930778   0.49130897
#>    29   47 0.509456264 0.8118322704 0.480817562 0.8315532035   0.48595433
#>    48   34 0.460583336 0.8479285424 0.510450198 0.8670427978   0.46459743
#>    72   34 0.503676928 0.8300930778 0.510450198 0.8670427978   0.49353904
#>    70   82 0.218576046 0.9523664380 0.230618606 0.9864357457   0.13863632
#>    78   46 0.200491547 0.1117229888 0.285892575 0.0901326655   0.25826035
#>    22   16 0.352466091 0.7119194799 0.560902853 0.7875257028   0.48052471
#>    17   66 0.050313407 0.0901488866 0.147822819 0.0219815501   0.12262261
#>    21   98 0.183239139 0.1291845698 0.133115972 0.1867342914   0.11391530
#>    27   58 0.481699473 0.2591508101 0.393377769 0.2399969693   0.44785412
#>     3   61 0.404849654 0.1054307950 0.471025348 0.1167078678   0.42702707
#>    59   48 0.412139647 0.8591150901 0.460583336 0.8479285424   0.44996330
#>    38   34 0.539587254 0.8894573287 0.510450198 0.8670427978   0.55454879
#>    62   70 0.177772755 0.9035000303 0.218576046 0.9523664380   0.13367750
#>    24   82 0.352022916 0.9199484307 0.230618606 0.9864357457   0.32674446
#>     5   43 0.534826596 0.1569388961 0.616367176 0.2589976039   0.54899936
#>    18   94 0.730792427 0.6181761825 0.442146235 0.4385407696   0.54720515
#>    43   65 0.616367176 0.2589976039 0.673283180 0.4553510055   0.65946412
#>    84   61 0.411972466 0.0902972213 0.471025348 0.1167078678   0.43782838
#>     3   46 0.404849654 0.1054307950 0.285892575 0.0901326655   0.34895906
#>    61   97 0.471025348 0.1167078678 0.499803189 0.1606287912   0.42845213
#>    97    5 0.499803189 0.1606287912 0.534826596 0.1569388961   0.52332986
#>    61    5 0.471025348 0.1167078678 0.534826596 0.1569388961   0.51327254
#>    16   38 0.560902853 0.7875257028 0.539587254 0.8894573287   0.55512850
#>    65   95 0.673283180 0.4553510055 0.704264732 0.4476675153   0.70928261
#>    46   66 0.285892575 0.0901326655 0.147822819 0.0219815501   0.29670418
#>    93   23 0.008134985 0.4799288952 0.028039999 0.6894806926  -6.82663985
#>    23   71 0.028039999 0.6894806926 0.053500904 0.8572904488  -0.70892009
#>    18   16 0.730792427 0.6181761825 0.560902853 0.7875257028   0.54243741
#>    81   28 0.463768778 0.9763151701 0.722729460 0.9861931293   0.47815905
#>   100   33 0.610946249 0.8777304193 0.635434889 0.9325374656   0.58184268
#>    76   63 0.692741934 0.2512649054 0.779288476 0.3933125038   0.69065325
#>    95   43 0.704264732 0.4476675153 0.616367176 0.2589976039   0.66430878
#>     5   87 0.534826596 0.1569388961 0.500872036 0.0398197218   0.55584669
#>    61   87 0.471025348 0.1167078678 0.500872036 0.0398197218   0.52925214
#>    46   13 0.285892575 0.0901326655 0.461919313 0.0000000000   0.34905479
#>   100   16 0.610946249 0.8777304193 0.560902853 0.7875257028   0.65470530
#>    13   87 0.461919313 0.0000000000 0.500872036 0.0398197218   0.44885894
#>    18   25 0.730792427 0.6181761825 0.829099338 0.5311600985   0.79594159
#>    33   81 0.635434889 0.9325374656 0.463768778 0.9763151701   0.55625867
#>    91   76 0.788355488 0.2992225554 0.692741934 0.2512649054   0.76732160
#>    76   43 0.692741934 0.2512649054 0.616367176 0.2589976039   0.64584028
#>    54   74 0.579290718 0.0002597358 0.627662914 0.0761813713   0.56244223
#>    39   15 0.803069258 0.6834260554 0.831945173 0.7418800242   0.78242277
#>    39   79 0.803069258 0.6834260554 0.733022914 0.7423791193   0.74045874
#>    63   91 0.779288476 0.3933125038 0.788355488 0.2992225554   0.79392460
#>    63   25 0.779288476 0.3933125038 0.829099338 0.5311600985   0.77869700
#>    13   54 0.461919313 0.0000000000 0.579290718 0.0002597358   0.52064621
#>   100   79 0.610946249 0.8777304193 0.733022914 0.7423791193   0.68628773
#>    39   18 0.803069258 0.6834260554 0.730792427 0.6181761825   0.79150245
#>    25   57 0.829099338 0.5311600985 0.842049523 0.3518000813   0.86357627
#>    76   74 0.692741934 0.2512649054 0.627662914 0.0761813713   0.70892103
#>    15   89 0.831945173 0.7418800242 0.778532654 0.8358990302   0.78265523
#>    28   92 0.722729460 0.9861931293 0.045690519 1.0000000000   0.47896820
#>    28   33 0.722729460 0.9861931293 0.635434889 0.9325374656   0.71400490
#>    33   89 0.635434889 0.9325374656 0.778532654 0.8358990302   0.71153124
#>    85   40 0.973025183 0.3713070080 0.957481846 0.4361286926   0.90346425
#>    56   57 0.845967915 0.2373956090 0.842049523 0.3518000813   0.84542565
#>    15   31 0.831945173 0.7418800242 0.861475836 0.7837586948   0.80773131
#>    41   64 0.827829212 0.8853814521 0.797973227 0.8978651909   0.80089743
#>    39   77 0.803069258 0.6834260554 0.851273284 0.6355161161   0.85821385
#>    76   30 0.692741934 0.2512649054 0.815340095 0.1555538674   0.76717359
#>    28   64 0.722729460 0.9861931293 0.797973227 0.8978651909   0.87471200
#>    31   41 0.861475836 0.7837586948 0.827829212 0.8853814521   0.83316663
#>    36   96 0.912395874 0.7032821670 0.914448054 0.8407344856   0.93153219
#>    25   40 0.829099338 0.5311600985 0.957481846 0.4361286926   0.93365211
#>     7   25 0.997348092 0.6268539543 0.829099338 0.5311600985   0.93514484
#>    56   85 0.845967915 0.2373956090 0.973025183 0.3713070080   0.91713640
#>    77   25 0.851273284 0.6355161161 0.829099338 0.5311600985   0.92046190
#>    35   56 1.000000000 0.1167906558 0.845967915 0.2373956090   0.91379703
#>    96   51 0.914448054 0.8407344856 0.988116029 0.9060811552   0.90396709
#>    96   41 0.914448054 0.8407344856 0.827829212 0.8853814521   0.85888342
#>    36    7 0.912395874 0.7032821670 0.997348092 0.6268539543   1.04773320
#>    85   67 0.973025183 0.3713070080 0.969617059 0.4397483955   0.97318924
#>    30   74 0.815340095 0.1555538674 0.627662914 0.0761813713   0.73253362
#>    26   35 0.831818580 0.0533768601 1.000000000 0.1167906558   0.90372758
#>    36   77 0.912395874 0.7032821670 0.851273284 0.6355161161   0.92430518
#>    51   28 0.988116029 0.9060811552 0.722729460 0.9861931293   0.88766827
#>     7   85 0.997348092 0.6268539543 0.973025183 0.3713070080   2.25366236
#>    35   85 1.000000000 0.1167906558 0.973025183 0.3713070080   0.97440455
#>     1   53 0.181984080 0.5734872478 0.257069450 0.4998742742   0.20446911
#>    52   73 0.194315905 0.4551471535 0.176442448 0.4719123770   0.21394872
#>    14   52 0.287310817 0.4642783664 0.194315905 0.4551471535   0.24120247
#>    53   86 0.257069450 0.4998742742 0.318888661 0.6340997645   0.26854045
#>    32   14 0.339546529 0.5463479760 0.287310817 0.4642783664   0.30548929
#>    73   75 0.176442448 0.4719123770 0.083136903 0.4255804275   0.10405024
#>     9    1 0.072727101 0.5894686361 0.181984080 0.5734872478   0.12407361
#>    73   42 0.176442448 0.4719123770 0.177346563 0.3690873285   0.14395962
#>     1    6 0.181984080 0.5734872478 0.206508258 0.6998410128   0.23979767
#>    14   69 0.287310817 0.4642783664 0.318899919 0.3235354913   0.30294091
#>    52   83 0.194315905 0.4551471535 0.241985792 0.3553199246   0.21890806
#>    75   11 0.083136903 0.4255804275 0.140260446 0.3026416731   0.10487340
#>    75    4 0.083136903 0.4255804275 0.066494593 0.4173131538   0.05372072
#>    75   93 0.083136903 0.4255804275 0.008134985 0.4799288952   0.07788620
#>    23    9 0.028039999 0.6894806926 0.072727101 0.5894686361  -0.01700321
#>     1   49 0.181984080 0.5734872478 0.061971719 0.6669235077   0.14259674
#>    22   86 0.352466091 0.7119194799 0.318888661 0.6340997645   0.28022742
#>    23   90 0.028039999 0.6894806926 0.065061541 0.5679794213  -0.07036239
#>    71    6 0.053500904 0.8572904488 0.206508258 0.6998410128   0.11572317
#>    14   94 0.287310817 0.4642783664 0.442146235 0.4385407696   0.36017742
#>    14   20 0.287310817 0.4642783664 0.375491760 0.3422080701   0.32540571
#>    16   32 0.560902853 0.7875257028 0.339546529 0.5463479760   0.50492362
#>    68   50 0.325025648 0.1960531145 0.391651093 0.2772765735   0.33481053
#>    83   45 0.241985792 0.3553199246 0.166294973 0.2430968603   0.20782904
#>     8   88 0.018250055 0.2988399633 0.013989242 0.2089244778   0.07086250
#>     8    4 0.018250055 0.2988399633 0.066494593 0.4173131538   0.07791978
#>    71   60 0.053500904 0.8572904488 0.203687760 0.8353304547   0.11681466
#>    44   19 0.338935752 0.8078522612 0.345653391 0.7770155566   0.26182639
#>    19   59 0.345653391 0.7770155566 0.412139647 0.8591150901   0.40537424
#>     8   11 0.018250055 0.2988399633 0.140260446 0.3026416731   0.08081083
#>    37   93 0.000000000 0.2540724224 0.008134985 0.4799288952  -0.23401688
#>    68   69 0.325025648 0.1960531145 0.318899919 0.3235354913   0.32968854
#>    80   50 0.485607825 0.3309353779 0.391651093 0.2772765735   0.42895958
#>    71   49 0.053500904 0.8572904488 0.061971719 0.6669235077   0.10209650
#>    27   99 0.481699473 0.2591508101 0.417725664 0.2638078999   0.45228025
#>     3   68 0.404849654 0.1054307950 0.325025648 0.1960531145   0.38889263
#>    60   10 0.203687760 0.8353304547 0.265439129 0.9374370246   0.27647806
#>    29   22 0.509456264 0.8118322704 0.352466091 0.7119194799   0.47040019
#>    43   27 0.616367176 0.2589976039 0.481699473 0.2591508101   0.54907019
#>    68   78 0.325025648 0.1960531145 0.200491547 0.1117229888   0.24790536
#>    21   45 0.183239139 0.1291845698 0.166294973 0.2430968603   0.23856371
#>    17   88 0.050313407 0.0901488866 0.013989242 0.2089244778   0.06651450
#>    19    2 0.345653391 0.7770155566 0.345190939 0.8258541414   0.39880818
#>    65   12 0.673283180 0.4553510055 0.570160514 0.4049472651   0.54796085
#>    60   62 0.203687760 0.8353304547 0.177772755 0.9035000303   0.23487320
#>    48   55 0.460583336 0.8479285424 0.464376280 0.8182976599   0.43070964
#>    16   94 0.560902853 0.7875257028 0.442146235 0.4385407696   0.51333339
#>     2   24 0.345190939 0.8258541414 0.352022916 0.9199484307   0.36411327
#>    47   72 0.480817562 0.8315532035 0.503676928 0.8300930778   0.49130897
#>    55   47 0.464376280 0.8182976599 0.480817562 0.8315532035   0.48595433
#>    81   34 0.463768778 0.9763151701 0.510450198 0.8670427978   0.46459743
#>    47   34 0.480817562 0.8315532035 0.510450198 0.8670427978   0.49353904
#>    92   82 0.045690519 1.0000000000 0.230618606 0.9864357457   0.13863632
#>    16   29 0.560902853 0.7875257028 0.509456264 0.8118322704   0.55203398
#>    21   66 0.183239139 0.1291845698 0.147822819 0.0219815501   0.12262261
#>    17   98 0.050313407 0.0901488866 0.133115972 0.1867342914   0.11391530
#>    17   37 0.050313407 0.0901488866 0.000000000 0.2540724224 -22.23135048
#>    97   58 0.499803189 0.1606287912 0.393377769 0.2399969693   0.44785412
#>    58   61 0.393377769 0.2399969693 0.471025348 0.1167078678   0.42702707
#>    81   48 0.463768778 0.9763151701 0.460583336 0.8479285424   0.44996330
#>    92   70 0.045690519 1.0000000000 0.218576046 0.9523664380   0.13367750
#>    10   82 0.265439129 0.9374370246 0.230618606 0.9864357457   0.32674446
#>    81   38 0.463768778 0.9763151701 0.539587254 0.8894573287   0.49049662
#>    65   94 0.673283180 0.4553510055 0.442146235 0.4385407696   0.54720515
#>    84   46 0.411972466 0.0902972213 0.285892575 0.0901326655   0.34895906
#>    43   80 0.616367176 0.2589976039 0.485607825 0.3309353779   0.56672387
#>    27    5 0.481699473 0.2591508101 0.534826596 0.1569388961   0.52332986
#>    81   59 0.463768778 0.9763151701 0.412139647 0.8591150901   0.41903449
#>    72   38 0.503676928 0.8300930778 0.539587254 0.8894573287   0.55512850
#>    18   95 0.730792427 0.6181761825 0.704264732 0.4476675153   0.70928261
#>    13   66 0.461919313 0.0000000000 0.147822819 0.0219815501   0.29670418
#>    92   23 0.045690519 1.0000000000 0.028039999 0.6894806926  -6.82663985
#>    92   71 0.045690519 1.0000000000 0.053500904 0.8572904488  -0.70892009
#>    92   62 0.045690519 1.0000000000 0.177772755 0.9035000303   0.09673609
#>    82   28 0.230618606 0.9864357457 0.722729460 0.9861931293   0.47815905
#>    81   24 0.463768778 0.9763151701 0.352022916 0.9199484307   0.35067483
#>    79   18 0.733022914 0.7423791193 0.730792427 0.6181761825   0.62525532
#>    38   33 0.539587254 0.8894573287 0.635434889 0.9325374656   0.58184268
#>    95   63 0.704264732 0.4476675153 0.779288476 0.3933125038   0.69065325
#>    74    5 0.627662914 0.0761813713 0.534826596 0.1569388961   0.62584036
#>    74   87 0.627662914 0.0761813713 0.500872036 0.0398197218   0.55584669
#>    87   84 0.500872036 0.0398197218 0.411972466 0.0902972213   0.45769222
#>    84   13 0.411972466 0.0902972213 0.461919313 0.0000000000   0.34905479
#>    79   16 0.733022914 0.7423791193 0.560902853 0.7875257028   0.65470530
#>    89  100 0.778532654 0.8358990302 0.610946249 0.8777304193   0.69841111
#>    56   76 0.845967915 0.2373956090 0.692741934 0.2512649054   0.76732160
#>    74   43 0.627662914 0.0761813713 0.616367176 0.2589976039   0.64584028
#>    79   15 0.733022914 0.7423791193 0.831945173 0.7418800242   0.78242277
#>    57   91 0.842049523 0.3518000813 0.788355488 0.2992225554   0.79392460
#>    95   25 0.704264732 0.4476675153 0.829099338 0.5311600985   0.77869700
#>    26   54 0.831818580 0.0533768601 0.579290718 0.0002597358   0.71858220
#>    87   54 0.500872036 0.0398197218 0.579290718 0.0002597358   0.52064621
#>    89   79 0.778532654 0.8358990302 0.733022914 0.7423791193   0.68628773
#>    36   39 0.912395874 0.7032821670 0.803069258 0.6834260554   0.85784672
#>    77   18 0.851273284 0.6355161161 0.730792427 0.6181761825   0.79150245
#>    40   57 0.957481846 0.4361286926 0.842049523 0.3518000813   0.86357627
#>    57   63 0.842049523 0.3518000813 0.779288476 0.3933125038   0.85729500
#>    64   33 0.797973227 0.8978651909 0.635434889 0.9325374656   0.71400490
#>    64   89 0.797973227 0.8978651909 0.778532654 0.8358990302   0.71153124
#>    89   31 0.778532654 0.8358990302 0.861475836 0.7837586948   0.80773131
#>    36   15 0.912395874 0.7032821670 0.831945173 0.7418800242   0.88015954
#>    56   30 0.845967915 0.2373956090 0.815340095 0.1555538674   0.76717359
#>    89   41 0.778532654 0.8358990302 0.827829212 0.8853814521   0.83316663
#>    31   96 0.861475836 0.7837586948 0.914448054 0.8407344856   0.93153219
#>    67   40 0.969617059 0.4397483955 0.957481846 0.4361286926   0.93365211
#>    67   25 0.969617059 0.4397483955 0.829099338 0.5311600985   0.93514484
#>    57   85 0.842049523 0.3518000813 0.973025183 0.3713070080   0.91713640
#>    41   51 0.827829212 0.8853814521 0.988116029 0.9060811552   0.90396709
#>    51    7 0.988116029 0.9060811552 0.997348092 0.6268539543   1.04773320
#>    26   74 0.831818580 0.0533768601 0.627662914 0.0761813713   0.73253362
#>    30   35 0.815340095 0.1555538674 1.000000000 0.1167906558   0.90372758
#>     7   77 0.997348092 0.6268539543 0.851273284 0.6355161161   0.92430518
#>    41   28 0.827829212 0.8853814521 0.722729460 0.9861931293   0.88766827
#>    51   36 0.988116029 0.9060811552 0.912395874 0.7032821670   1.04294310
#>    53   32 0.257069450 0.4998742742 0.339546529 0.5463479760   0.26854045
#>    73   90 0.176442448 0.4719123770 0.065061541 0.5679794213   0.10405024
#>    11   83 0.140260446 0.3026416731 0.241985792 0.3553199246   0.19924827
#>    22    6 0.352466091 0.7119194799 0.206508258 0.6998410128   0.28022742
#>     8   37 0.018250055 0.2988399633 0.000000000 0.2540724224   0.07086250
#>    55   19 0.464376280 0.8182976599 0.345653391 0.7770155566   0.42089549
#>    20   80 0.375491760 0.3422080701 0.485607825 0.3309353779   0.43425587
#>     2   60 0.345190939 0.8258541414 0.203687760 0.8353304547   0.27507800
#>     3   58 0.404849654 0.1054307950 0.393377769 0.2399969693   0.38889263
#>    29   55 0.509456264 0.8118322704 0.464376280 0.8182976599   0.47040019
#>    68   21 0.325025648 0.1960531145 0.183239139 0.1291845698   0.24790536
#>    16   72 0.560902853 0.7875257028 0.503676928 0.8300930778   0.55203398
#>    21   17 0.183239139 0.1291845698 0.050313407 0.0901488866   0.12262261
#>    97   27 0.499803189 0.1606287912 0.481699473 0.2591508101   0.44785412
#>    65   18 0.673283180 0.4553510055 0.730792427 0.6181761825   0.54720515
#>    12   43 0.570160514 0.4049472651 0.616367176 0.2589976039   0.65946412
#>     3   84 0.404849654 0.1054307950 0.411972466 0.0902972213   0.43782838
#>    92   93 0.045690519 1.0000000000 0.008134985 0.4799288952  -6.82663985
#>    82   81 0.230618606 0.9864357457 0.463768778 0.9763151701   0.47815905
#>    38  100 0.539587254 0.8894573287 0.610946249 0.8777304193   0.58184268
#>    95   76 0.704264732 0.4476675153 0.692741934 0.2512649054   0.69065325
#>    56   91 0.845967915 0.2373956090 0.788355488 0.2992225554   0.76732160
#>    36   31 0.912395874 0.7032821670 0.861475836 0.7837586948   0.88015954
#>    67    7 0.969617059 0.4397483955 0.997348092 0.6268539543   0.93514484
#>    26   30 0.831818580 0.0533768601 0.815340095 0.1555538674   0.73253362
#>    35    7 1.000000000 0.1167906558 0.997348092 0.6268539543   2.25366236
#>           my1           mx2          my2 bp1 bp2
#>   0.521321926   0.213948723  0.493987848   0   0
#>   0.455749996   0.213948723  0.493987848   0   0
#>   0.575939711   0.457534176  0.620431197   0   0
#>   0.510366431   0.241202466  0.455749996   0   0
#>   0.500581745   0.077886197  0.497260623   0   0
#>   0.525622187   0.204469111  0.521321926   0   0
#>   0.559040659   0.125647844  0.525622187   0   0
#>   0.420210265   0.109786888  0.363222738   0   0
#>   0.420217234   0.218908062  0.405595127   0   0
#>   0.418010765   0.302940907  0.393870016   0   0
#>   0.627822976   0.280227421  0.696935011   0   0
#>   0.579279101   0.239797672  0.627822976   0   0
#>   0.393870016   0.250813598  0.267728821   0   0
#>   0.405595127   0.199248267  0.313290577   0   0
#>   0.360939685   0.077919776  0.343600960   0   0
#>   0.463911810  -0.028755889  0.387041320   0   0
#>   0.637669872   0.006109628  0.619692317   0   0
#>   0.497260623  -0.070362389  0.593106517   0   0
#>   0.609365105   0.124073613  0.559040659   0   0
#>   0.646688840   0.115723169  0.764687237   0   0
#>   0.363222738   0.199248267  0.313290577   0   0
#>   0.470628465   0.513333387  0.609014787   0   0
#>   0.267728821   0.256650161  0.256655928   0   0
#>   0.768713719   0.276918743  0.736917525   0   0
#>   0.424030475   0.434255871  0.372774150   0   0
#>   0.398912139   0.357008630  0.303131708   0   0
#>   0.616733394   0.457534176  0.620431197   0   0
#>   0.255964105   0.398686339  0.258922635   0   0
#>   0.296720514   0.094737148  0.247273825   0   0
#>   0.251288128 -22.231350475 -6.659127517   0   0
#>   0.736917525   0.420895487  0.751985565   0   0
#>   0.319371903   0.398686339  0.258922635   0   0
#>   0.303131708   0.428959585  0.321037955   0   0
#>   0.765747930   0.259283510  0.768713719   0   0
#>   0.774904272   0.259283510  0.768713719   0   0
#>   0.801812907   0.275078003  0.840128872   0   0
#>   0.372774150   0.500303107  0.399450489   0   0
#>   0.258922635   0.446004180  0.210537721   0   0
#>   0.840128872   0.261826390  0.774904272   0   0
#>   0.796623004   0.430709644  0.829046316   0   0
#>   0.250816708   0.082898737  0.248003539   0   0
#>   0.375576073  -0.028755889  0.387041320   0   0
#>   0.260165538   0.357008630  0.303131708   0   0
#>   0.399450489   0.566723867  0.323570059   0   0
#>   0.188969114   0.094737148  0.247273825   0   0
#>   0.248003539   0.094737148  0.247273825   0   0
#>   0.764080887   0.006109628  0.619692317   0   0
#>   0.845763561   0.096736087  0.931225181   0   0
#>   0.296751237   0.429911069  0.319371903   0   0
#>   0.171842526   0.334810532  0.255964105   0   0
#>   0.861034884   0.298201241  0.876561131   0   0
#>   0.699906454   0.420895487  0.751985565   0   0
#>   0.291481453   0.452280249  0.296751237   0   0
#>   0.175822494   0.151231835  0.080307022   0   0
#>   0.195630307   0.256650161  0.256655928   0   0
#>   0.160045689   0.082898737  0.248003539   0   0
#>   0.801940361   0.364113265  0.871775405   0   0
#>   0.581059171   0.500303107  0.399450489   0   0
#>   0.886196405   0.228527612  0.902588402   0   0
#>   0.129871808   0.258260354  0.160530812   0   0
#>   0.080307022   0.226770910  0.035973580   0   0
#>   0.833521059   0.485370320  0.857867733   0   0
#>   0.902588402   0.247260438  0.961390307   0   0
#>   0.871775405   0.419034494  0.926049669   0   0
#>   0.876561131   0.326744460  1.017874950   0   0
#>   0.816133812   0.552033982  0.835352580   0   0
#>   0.808357805   0.491308975  0.816133812   0   0
#>   0.912061779   0.485370320  0.857867733   0   0
#>   0.851047124   0.554548786  0.839863401   0   0
#>   0.999785934   0.247260438  0.961390307   0   0
#>   0.160530812   0.226770910  0.035973580   0   0
#>   0.683998086   0.504923615  0.616733394   0   0
#>   0.089758560  -4.693199157 -6.798994485   0   1
#>   0.119409034   0.193781823  0.188969114   0   0
#>   0.202007310   0.446004180  0.210537721   0   0
#>   0.175093520   0.437828383  0.111709652   0   0
#>   0.912424871   0.430709644  0.829046316   0   0
#>   0.839863401   0.490496621  0.923125976   0   0
#>   0.981787936   0.228527612  0.902588402   0   0
#>   1.017874950   0.350674828  1.061571253   0   0
#>   0.229218546   0.625840361  0.167825844   0   0
#>   0.591449911   0.542437407  0.599110935   0   0
#>   0.352930987   0.664000247  0.351616121   0   0
#>   0.111709652   0.457692221  0.067295107   0   0
#>   0.069882153   0.341244263  0.129871808   0   0
#>   0.175991026   0.513272544  0.120415047   0   0
#>   0.215876204   0.513272544  0.120415047   0   0
#>   0.120415047   0.529252136  0.095073465   0   0
#>   0.839512725   0.568481479  0.842305056   0   0
#>   0.534204754   0.664000247  0.351616121   0   0
#>  -0.105706699   0.226770910  0.035973580   0   0
#>   1.234875284  -0.070362389  0.593106517   0   0
#>   0.887132231   0.102096502  0.764080887   0   0
#>   0.599110935   0.625255324  0.682192958   0   0
#>   3.998456503   0.588460145  1.106801946   0   0
#>   0.923608821   0.698411106  0.871524238   0   0
#>   0.349926769   0.708562392  0.339015112   0   0
#>   0.351472380   0.664000247  0.351616121   0   0
#>   0.087363315   0.529252136  0.095073465   0   0
#>   0.095073465   0.457692221  0.067295107   0   0
#>  -0.003467317   0.296704176 -0.105706699   0   0
#>   0.794470163   0.568481479  0.842305056   0   0
#>   0.051738173   0.520646211 -0.018486070   0   0
#>   0.592739386   0.739911849  0.529439466   0   0
#>   0.980529872   0.588460145  1.106801946   0   0
#>   0.221866402   0.708562392  0.339015112   0   0
#>   0.169061587   0.664308784  0.351472380   0   0
#>   0.064365048   0.718582203 -0.035116911   0   0
#>   0.729984549   0.857846725  0.692725561   0   0
#>   0.680124087   0.782422768  0.729984549   0   0
#>   0.347241072   0.708562392  0.339015112   0   0
#>   0.471449542   0.857295000  0.443048362   0   0
#>  -0.018486070   0.539500278 -8.538407082   0   1
#>   0.822955129   0.654705303  0.794470163   0   0
#>   0.623583319   0.740458736  0.680124087   0   0
#>   0.443501884   0.857295000  0.443048362   0   0
#>   0.145614303   0.645840278  0.169061587   0   0
#>   0.776059661   0.807731307  0.790305465   0   0
#>   5.639696282   0.509076584  7.116102198   0   1
#>   0.902548098   0.588460145  1.106801946   0   0
#>   0.890951924   0.698411106  0.871524238   0   0
#>   0.388901645   0.973189242  0.405620727   0   0
#>   0.294646375   0.917136404  0.297102493   0   0
#>   0.790305465   0.880159536  0.739232840   0   0
#>   0.862915170   0.874712005  1.039449377   0   0
#>   0.690704211   0.791502451  0.623583319   0   0
#>   0.220231159   0.708921027  0.145614303   0   0
#>   1.039449377   0.714004902  0.902548098   0   0
#>   0.830767171   0.858883418  0.839281828   0   0
#>   0.771737938   1.042943097  0.770074560   0   0
#>   0.538170711   0.863576269  0.443501884   0   0
#>   0.540465341   0.920461903  0.566280857   0   0
#>   0.297102493   0.974404548  0.242765566   0   0
#>   0.566280857   0.795941594  0.592739386   0   0
#>   0.165359935   0.974404548  0.242765566   0   0
#>   0.926747907   1.042943097  0.770074560   0   0
#>   0.839281828   0.903967085  0.926747907   0   0
#>   0.768286052   0.924305180  0.631092156   0   0
#>   0.405620727   1.625722294  0.438114422   0   0
#>   0.089782064   0.708921027  0.145614303   0   0
#>   0.117391185   2.866620176 -5.088448630   0   1
#>   0.631092156   0.858213846  0.690704211   0   0
#>   1.052956743   1.930141937  4.506354477   0   1
#>   0.378347207   1.625722294  0.438114422   0   0
#>   0.242765566   2.253662363  0.378347207   0   0
#>   0.521321926   0.261289767  0.579279101   0   0
#>   0.493987848   0.144752214  0.420217234   0   0
#>   0.455749996   0.244908099  0.418010765   0   0
#>   0.575939711   0.261289767  0.579279101   0   0
#>   0.510366431   0.367923197  0.470628465   0   0
#>   0.500581745   0.143959625  0.420210265   0   0
#>   0.559040659   0.135574974  0.637669872   0   0
#>   0.420210265   0.144752214  0.420217234   0   0
#>   0.627822976   0.142596740  0.646688840   0   0
#>   0.393870016   0.325405705  0.398912139   0   0
#>   0.405595127   0.244908099  0.418010765   0   0
#>   0.360939685   0.109786888  0.363222738   0   0
#>   0.463911810   0.104873402  0.360939685   0   0
#>   0.497260623   0.053720723  0.463911810   0   0
#>   0.609365105   0.006109628  0.619692317   0   0
#>   0.646688840   0.135574974  0.637669872   0   0
#>   0.696935011   0.457534176  0.620431197   0   0
#>   0.593106517  -0.017003209  0.609365105   0   0
#>   0.764687237   0.116814656  0.765747930   0   0
#>   0.424030475   0.367923197  0.470628465   0   0
#>   0.398912139   0.360177422  0.424030475   0   0
#>   0.616733394   0.513333387  0.609014787   0   0
#>   0.255964105   0.329688541  0.260165538   0   0
#>   0.296720514   0.250813598  0.267728821   0   0
#>   0.251288128   0.080810831  0.250816708   0   0
#>   0.343600960  -0.028755889  0.387041320   0   0
#>   0.765747930   0.128514367  0.845763561   0   0
#>   0.774904272   0.385347997  0.801812907   0   0
#>   0.796623004   0.398808182  0.801940361   0   0
#>   0.250816708   0.077919776  0.343600960   0   0
#>   0.375576073  -4.655870925  0.534844113   0   1
#>   0.260165538   0.256650161  0.256655928   0   0
#>   0.321037955   0.429911069  0.319371903   0   0
#>   0.764080887   0.115723169  0.764687237   0   0
#>   0.296751237   0.446004180  0.210537721   0   0
#>   0.171842526   0.341244263  0.129871808   0   0
#>   0.861034884   0.234873205  0.886196405   0   0
#>   0.699906454   0.480524708  0.683998086   0   0
#>   0.291481453   0.548999359  0.229218546   0   0
#>   0.175822494   0.258260354  0.160530812   0   0
#>   0.195630307   0.193781823  0.188969114   0   0
#>   0.160045689 -22.231350475 -6.659127517   0   0
#>   0.801940361   0.385347997  0.801812907   0   0
#>   0.581059171   0.659464117  0.352930987   0   0
#>   0.886196405   0.128514367  0.845763561   0   0
#>   0.829046316   0.465666822  0.833521059   0   0
#>   0.609014787   0.542437407  0.599110935   0   0
#>   0.871775405   0.298201241  0.876561131   0   0
#>   0.816133812   0.493539038  0.851047124   0   0
#>   0.808357805   0.465666822  0.833521059   0   0
#>   0.912061779   0.490496621  0.923125976   0   0
#>   0.851047124   0.485370320  0.857867733   0   0
#>   0.999785934   0.478968204  5.639696282   0   0
#>   0.835352580   0.480524708  0.683998086   0   0
#>   0.089758560   0.151231835  0.080307022   0   0
#>   0.119409034   0.066514496  0.160045689   0   0
#>  -6.659127517 -27.806532547 -8.370330281   0   1
#>   0.202007310   0.428452132  0.175991026   0   0
#>   0.175093520   0.428452132  0.175991026   0   0
#>   0.912424871   0.464597430  0.912061779   0   0
#>   0.981787936   0.138636322  0.999785934   0   0
#>   1.017874950   0.247260438  0.961390307   0   0
#>   0.923125976   0.556258672  0.980529872   0   0
#>   0.591449911   0.547960854  0.581059171   0   0
#>   0.069882153   0.349054792 -0.003467317   0   0
#>   0.323570059   0.549070193  0.291481453   0   0
#>   0.215876204   0.548999359  0.229218546   0   0
#>   0.926049669   0.449963301  0.912424871   0   0
#>   0.839512725   0.554548786  0.839863401   0   0
#>   0.534204754   0.739911849  0.529439466   0   0
#>  -0.105706699   0.074981408 -3.273924778   0   1
#>   1.234875284  -0.708920086  0.887132231   0   0
#>   0.887132231   0.096736087  0.931225181   0   0
#>   0.931225181   0.133677498  0.981787936   0   0
#>   3.998456503   0.478968204  5.639696282   0   0
#>   1.061571253   0.419034494  0.926049669   0   0
#>   0.682192958   0.740458736  0.680124087   0   0
#>   0.923608821   0.556258672  0.980529872   0   0
#>   0.349926769   0.778696995  0.471449542   0   0
#>   0.167825844   0.555846690  0.087363315   0   0
#>   0.087363315   0.562442235  0.064365048   0   0
#>   0.067295107   0.448858937  0.051738173   0   0
#>  -0.003467317   0.448858937  0.051738173   0   0
#>   0.794470163   0.625255324  0.682192958   0   0
#>   0.871524238   0.686287728  0.822955129   0   0
#>   0.221866402   0.767173588  0.220231159   0   0
#>   0.169061587   0.625840361  0.167825844   0   0
#>   0.729984549   0.782655232  0.776059661   0   0
#>   0.347241072   0.845425645  0.294646375   0   0
#>   0.471449542   0.739911849  0.529439466   0   0
#>  -0.035116911   1.516235429 -3.827296302   0   1
#>  -0.018486070   0.562442235  0.064365048   0   0
#>   0.822955129   0.782655232  0.776059661   0   0
#>   0.692725561   0.858213846  0.690704211   0   0
#>   0.623583319   0.795941594  0.592739386   0   0
#>   0.443501884   0.903464250  0.388901645   0   0
#>   0.443048362   0.793924599  0.347241072   0   0
#>   0.902548098   0.711531235  0.890951924   0   0
#>   0.890951924   0.800897426  0.862915170   0   0
#>   0.790305465   0.833166634  0.830767171   0   0
#>   0.739232840   0.857846725  0.692725561   0   0
#>   0.220231159   0.913797030  0.165359935   0   0
#>   0.830767171   0.800897426  0.862915170   0   0
#>   0.771737938   0.858883418  0.839281828   0   0
#>   0.538170711   0.973189242  0.405620727   0   0
#>   0.540465341   0.933652106  0.538170711   0   0
#>   0.297102493   0.903464250  0.388901645   0   0
#>   0.926747907   0.887668269  1.052956743   0   0
#>   0.768286052   4.625135529  0.886565375   0   1
#>   0.089782064   0.718582203 -0.035116911   0   0
#>   0.117391185   0.913797030  0.165359935   0   0
#>   0.631092156   0.920461903  0.566280857   0   0
#>   1.052956743   0.874712005  1.039449377   0   0
#>   0.770074560   1.047733204  0.768286052   0   0
#>   0.575939711   0.305489295  0.510366431   0   0
#>   0.500581745   0.125647844  0.525622187   0   0
#>   0.313290577   0.207829039  0.296720514   0   0
#>   0.696935011   0.276918743  0.736917525   0   0
#>   0.251288128  -0.234016884  0.375576073   0   0
#>   0.751985565   0.405374239  0.796623004   0   0
#>   0.372774150   0.428959585  0.321037955   0   0
#>   0.840128872   0.276478056  0.861034884   0   0
#>   0.171842526   0.427027069  0.175093520   0   0
#>   0.699906454   0.485954328  0.808357805   0   0
#>   0.175822494   0.238563715  0.195630307   0   0
#>   0.835352580   0.555128497  0.839512725   0   0
#>   0.089758560   0.113915295  0.119409034   0   0
#>   0.202007310   0.523329865  0.215876204   0   0
#>   0.591449911   0.709282613  0.534204754   0   0
#>   0.352930987   0.566723867  0.323570059   0   0
#>   0.111709652   0.348959058  0.069882153   0   0
#>   1.234875284  -8.739479122  1.373005819   0   1
#>   3.998456503   0.350674828  1.061571253   0   0
#>   0.923608821   0.568481479  0.842305056   0   0
#>   0.349926769   0.664308784  0.351472380   0   0
#>   0.221866402   0.845425645  0.294646375   0   0
#>   0.739232840   0.931532189  0.771737938   0   0
#>   0.540465341   1.625722294  0.438114422   0   0
#>   0.089782064   0.903727584  0.117391185   0   0
#>   0.378347207   4.214150351  0.388540125   0   1
#> 
#> $x
#>               [,1]         [,2]
#>   [1,] 0.181984080 0.5734872478
#>   [2,] 0.345190939 0.8258541414
#>   [3,] 0.404849654 0.1054307950
#>   [4,] 0.066494593 0.4173131538
#>   [5,] 0.534826596 0.1569388961
#>   [6,] 0.206508258 0.6998410128
#>   [7,] 0.997348092 0.6268539543
#>   [8,] 0.018250055 0.2988399633
#>   [9,] 0.072727101 0.5894686361
#>  [10,] 0.265439129 0.9374370246
#>  [11,] 0.140260446 0.3026416731
#>  [12,] 0.570160514 0.4049472651
#>  [13,] 0.461919313 0.0000000000
#>  [14,] 0.287310817 0.4642783664
#>  [15,] 0.831945173 0.7418800242
#>  [16,] 0.560902853 0.7875257028
#>  [17,] 0.050313407 0.0901488866
#>  [18,] 0.730792427 0.6181761825
#>  [19,] 0.345653391 0.7770155566
#>  [20,] 0.375491760 0.3422080701
#>  [21,] 0.183239139 0.1291845698
#>  [22,] 0.352466091 0.7119194799
#>  [23,] 0.028039999 0.6894806926
#>  [24,] 0.352022916 0.9199484307
#>  [25,] 0.829099338 0.5311600985
#>  [26,] 0.831818580 0.0533768601
#>  [27,] 0.481699473 0.2591508101
#>  [28,] 0.722729460 0.9861931293
#>  [29,] 0.509456264 0.8118322704
#>  [30,] 0.815340095 0.1555538674
#>  [31,] 0.861475836 0.7837586948
#>  [32,] 0.339546529 0.5463479760
#>  [33,] 0.635434889 0.9325374656
#>  [34,] 0.510450198 0.8670427978
#>  [35,] 1.000000000 0.1167906558
#>  [36,] 0.912395874 0.7032821670
#>  [37,] 0.000000000 0.2540724224
#>  [38,] 0.539587254 0.8894573287
#>  [39,] 0.803069258 0.6834260554
#>  [40,] 0.957481846 0.4361286926
#>  [41,] 0.827829212 0.8853814521
#>  [42,] 0.177346563 0.3690873285
#>  [43,] 0.616367176 0.2589976039
#>  [44,] 0.338935752 0.8078522612
#>  [45,] 0.166294973 0.2430968603
#>  [46,] 0.285892575 0.0901326655
#>  [47,] 0.480817562 0.8315532035
#>  [48,] 0.460583336 0.8479285424
#>  [49,] 0.061971719 0.6669235077
#>  [50,] 0.391651093 0.2772765735
#>  [51,] 0.988116029 0.9060811552
#>  [52,] 0.194315905 0.4551471535
#>  [53,] 0.257069450 0.4998742742
#>  [54,] 0.579290718 0.0002597358
#>  [55,] 0.464376280 0.8182976599
#>  [56,] 0.845967915 0.2373956090
#>  [57,] 0.842049523 0.3518000813
#>  [58,] 0.393377769 0.2399969693
#>  [59,] 0.412139647 0.8591150901
#>  [60,] 0.203687760 0.8353304547
#>  [61,] 0.471025348 0.1167078678
#>  [62,] 0.177772755 0.9035000303
#>  [63,] 0.779288476 0.3933125038
#>  [64,] 0.797973227 0.8978651909
#>  [65,] 0.673283180 0.4553510055
#>  [66,] 0.147822819 0.0219815501
#>  [67,] 0.969617059 0.4397483955
#>  [68,] 0.325025648 0.1960531145
#>  [69,] 0.318899919 0.3235354913
#>  [70,] 0.218576046 0.9523664380
#>  [71,] 0.053500904 0.8572904488
#>  [72,] 0.503676928 0.8300930778
#>  [73,] 0.176442448 0.4719123770
#>  [74,] 0.627662914 0.0761813713
#>  [75,] 0.083136903 0.4255804275
#>  [76,] 0.692741934 0.2512649054
#>  [77,] 0.851273284 0.6355161161
#>  [78,] 0.200491547 0.1117229888
#>  [79,] 0.733022914 0.7423791193
#>  [80,] 0.485607825 0.3309353779
#>  [81,] 0.463768778 0.9763151701
#>  [82,] 0.230618606 0.9864357457
#>  [83,] 0.241985792 0.3553199246
#>  [84,] 0.411972466 0.0902972213
#>  [85,] 0.973025183 0.3713070080
#>  [86,] 0.318888661 0.6340997645
#>  [87,] 0.500872036 0.0398197218
#>  [88,] 0.013989242 0.2089244778
#>  [89,] 0.778532654 0.8358990302
#>  [90,] 0.065061541 0.5679794213
#>  [91,] 0.788355488 0.2992225554
#>  [92,] 0.045690519 1.0000000000
#>  [93,] 0.008134985 0.4799288952
#>  [94,] 0.442146235 0.4385407696
#>  [95,] 0.704264732 0.4476675153
#>  [96,] 0.914448054 0.8407344856
#>  [97,] 0.499803189 0.1606287912
#>  [98,] 0.133115972 0.1867342914
#>  [99,] 0.417725664 0.2638078999
#> [100,] 0.610946249 0.8777304193
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,73,53), [2,13,7]
#> 2: (52,53,73), [1,10,3]
#> 3: (14,53,52), [2,11,5]
#> 4: (53,32,86), [29,13,5]
#> 5: (32,53,14), [3,26,4]
#> 6: (73,90,75), [19,9,7]
#> 7: (90,73,1), [1,8,6]
#> 8: (9,90,1), [7,18,21]
#> 9: (73,75,42), [24,10,6]
#> 10: (73,42,52), [15,2,9]
#> 11: (52,83,14), [14,3,15]
#> 12: (1,86,6), [28,22,13]
#> 13: (53,86,1), [12,1,4]
#> 14: (14,83,69), [27,34,11]
#> 15: (52,42,83), [23,11,10]
#> 16: (75,4,11), [39,24,17]
#> 17: (75,93,4), [25,16,19]
#> 18: (1,49,9), [20,8,22]
#> 19: (75,90,93), [30,17,6]
#> 20: (23,9,49), [18,60,21]
#> 21: (23,90,9), [8,20,30]
#> 22: (1,6,49), [31,18,12]
#> 23: (11,83,42), [15,24,37]
#> 24: (75,11,42), [23,9,16]
#> 25: (8,4,93), [17,52,39]
#> 26: (14,94,32), [78,5,33]
#> 27: (83,45,69), [58,14,37]
#> 28: (22,6,86), [12,29,40]
#> 29: (22,86,32), [4,35,28]
#> 30: (23,93,90), [19,21,116]
#> 31: (71,49,6), [22,44,60]
#> 32: (60,6,19), [40,45,44]
#> 33: (14,20,94), [47,26,34]
#> 34: (14,69,20), [43,33,14]
#> 35: (16,22,32), [29,78,91]
#> 36: (68,58,50), [48,53,64]
#> 37: (83,11,45), [57,27,23]
#> 38: (8,37,88), [94,51,52]
#> 39: (8,11,4), [16,25,51]
#> 40: (6,22,19), [41,32,28]
#> 41: (55,19,22), [40,66,50]
#> 42: (80,50,99), [48,62,55]
#> 43: (69,50,20), [55,34,53]
#> 44: (71,6,60), [32,61,31]
#> 45: (44,60,19), [32,46,49]
#> 46: (19,2,44), [49,45,71]
#> 47: (20,80,94), [54,33,55]
#> 48: (50,58,99), [63,42,36]
#> 49: (2,60,44), [45,46,65]
#> 50: (19,55,59), [76,71,41]
#> 51: (8,88,11), [59,39,38]
#> 52: (37,8,93), [25,0,38]
#> 53: (68,50,69), [43,58,36]
#> 54: (94,80,12), [108,72,47]
#> 55: (80,20,50), [43,42,47]
#> 56: (21,45,98), [57,93,69]
#> 57: (98,45,11), [37,59,56]
#> 58: (68,69,45), [27,69,53]
#> 59: (88,98,11), [57,51,70]
#> 60: (71,23,49), [20,31,117]
#> 61: (60,62,71), [118,44,73]
#> 62: (27,80,99), [42,63,67]
#> 63: (27,99,58), [48,95,62]
#> 64: (3,58,68), [36,74,96]
#> 65: (60,2,10), [81,73,49]
#> 66: (29,55,22), [41,91,83]
#> 67: (43,80,27), [62,102,108]
#> 68: (68,21,78), [75,89,69]
#> 69: (21,68,45), [58,56,68]
#> 70: (17,98,88), [59,94,93]
#> 71: (19,59,2), [80,46,50]
#> 72: (65,94,12), [54,104,103]
#> 73: (60,10,62), [79,61,65]
#> 74: (3,68,46), [89,106,64]
#> 75: (21,66,78), [115,68,92]
#> 76: (48,59,55), [50,77,97]
#> 77: (55,47,48), [86,76,83]
#> 78: (16,32,94), [26,119,35]
#> 79: (62,10,70), [88,99,73]
#> 80: (2,59,24), [111,81,71]
#> 81: (2,24,10), [100,65,80]
#> 82: (47,29,72), [90,85,83]
#> 83: (55,29,47), [82,77,66]
#> 84: (81,48,34), [86,101,97]
#> 85: (47,72,34), [98,86,82]
#> 86: (47,34,48), [84,77,85]
#> 87: (92,70,82), [88,156,99]
#> 88: (10,82,70), [87,79,100]
#> 89: (68,78,46), [115,74,68]
#> 90: (16,72,29), [82,91,112]
#> 91: (29,22,16), [35,90,66]
#> 92: (21,17,66), [0,75,93]
#> 93: (17,21,98), [56,70,92]
#> 94: (17,88,37), [38,0,70]
#> 95: (97,27,58), [63,107,109]
#> 96: (58,3,61), [105,107,64]
#> 97: (81,59,48), [76,84,111]
#> 98: (72,38,34), [101,85,112]
#> 99: (92,62,70), [79,87,118]
#> 100: (10,24,82), [122,88,81]
#> 101: (81,34,38), [98,137,84]
#> 102: (27,5,43), [128,67,109]
#> 103: (65,18,94), [119,72,113]
#> 104: (12,43,65), [120,72,108]
#> 105: (3,84,61), [131,96,106]
#> 106: (84,3,46), [74,132,105]
#> 107: (58,61,97), [110,95,96]
#> 108: (43,12,80), [54,67,104]
#> 109: (27,97,5), [110,102,95]
#> 110: (97,61,5), [130,109,107]
#> 111: (81,24,59), [80,97,122]
#> 112: (72,16,38), [125,98,90]
#> 113: (18,65,95), [120,136,103]
#> 114: (13,46,66), [115,0,132]
#> 115: (78,66,46), [114,89,75]
#> 116: (92,93,23), [30,117,0]
#> 117: (92,23,71), [60,118,116]
#> 118: (92,71,62), [61,99,117]
#> 119: (94,18,16), [123,78,103]
#> 120: (95,65,43), [104,127,113]
#> 121: (82,81,28), [157,156,122]
#> 122: (81,82,24), [100,111,121]
#> 123: (79,16,18), [119,143,133]
#> 124: (38,100,33), [138,137,125]
#> 125: (100,38,16), [112,133,124]
#> 126: (95,76,63), [145,146,127]
#> 127: (76,95,43), [120,140,126]
#> 128: (74,43,5), [102,129,140]
#> 129: (74,5,87), [130,141,128]
#> 130: (5,61,87), [131,129,110]
#> 131: (87,61,84), [105,134,130]
#> 132: (84,46,13), [114,134,106]
#> 133: (79,100,16), [125,123,149]
#> 134: (84,13,87), [148,131,132]
#> 135: (77,18,25), [136,173,151]
#> 136: (95,25,18), [135,113,146]
#> 137: (38,33,81), [157,101,124]
#> 138: (89,33,100), [124,149,159]
#> 139: (56,91,76), [145,166,161]
#> 140: (74,76,43), [127,128,154]
#> 141: (87,54,74), [147,129,148]
#> 142: (79,39,15), [150,155,143]
#> 143: (18,39,79), [142,123,151]
#> 144: (57,63,91), [145,161,153]
#> 145: (76,91,63), [144,126,139]
#> 146: (95,63,25), [153,136,126]
#> 147: (26,74,54), [141,0,179]
#> 148: (87,13,54), [0,141,134]
#> 149: (89,100,79), [133,155,138]
#> 150: (36,15,39), [142,165,164]
#> 151: (77,39,18), [143,135,165]
#> 152: (40,25,57), [153,160,170]
#> 153: (57,25,63), [146,144,152]
#> 154: (30,76,74), [140,179,166]
#> 155: (79,15,89), [162,149,142]
#> 156: (82,28,92), [0,87,121]
#> 157: (28,81,33), [137,158,121]
#> 158: (64,28,33), [157,159,167]
#> 159: (64,33,89), [138,163,158]
#> 160: (57,85,40), [178,152,172]
#> 161: (91,56,57), [172,144,139]
#> 162: (89,15,31), [164,168,155]
#> 163: (89,41,64), [167,159,168]
#> 164: (36,31,15), [162,150,169]
#> 165: (36,39,77), [151,181,150]
#> 166: (56,76,30), [154,174,139]
#> 167: (41,28,64), [158,163,183]
#> 168: (89,31,41), [176,163,162]
#> 169: (31,36,96), [184,176,164]
#> 170: (67,25,40), [152,178,171]
#> 171: (67,7,25), [173,170,182]
#> 172: (57,56,85), [186,160,161]
#> 173: (7,77,25), [135,171,181]
#> 174: (30,35,56), [186,166,180]
#> 175: (41,96,51), [184,183,176]
#> 176: (31,96,41), [175,168,169]
#> 177: (51,36,7), [181,0,184]
#> 178: (40,85,67), [182,170,160]
#> 179: (26,30,74), [154,147,180]
#> 180: (30,26,35), [0,174,179]
#> 181: (7,36,77), [165,173,177]
#> 182: (7,67,85), [178,185,171]
#> 183: (41,51,28), [0,167,175]
#> 184: (51,96,36), [169,177,175]
#> 185: (35,7,85), [182,186,0]
#> 186: (56,35,85), [185,172,174]
#> boundary nodes:  93 37 17 66 13 54 26 35 7 51 28 92 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.08533804 0.07706181 0.09015056 0.04670770 0.14354160 0.10172593 0.11705219 
#>                                                                              
#> 0.10984958 0.08771689 0.11800974 0.13019703 0.14972221 0.08322279 0.06608912 
#>                                                                              
#> 0.13634863 0.08559541 0.07819805 0.10485003 0.02281550 0.14823756 0.07609471 
#>                                                                              
#> 0.14882572 0.17250694 0.15911404 0.11714439 0.05959280 0.16607480 0.03731957 
#>                                                                              
#> 0.06498755 0.04726559 0.06545160 0.02934772 0.06691206 0.13551880 0.15347595 
#>                                                                              
#> 0.01905768 0.11605098 0.03405555 0.13801112 0.06629275 0.15724926 0.18137121 
#>                                                                              
#> 0.08621265 0.11236955 0.06540325 0.11612736 0.04074541 0.13258514 0.09546772 
#>                                                                              
#> 0.08125931 0.13715353 0.15440274 0.07189089 0.02454694 0.16555525 0.12117583 
#>                                                                              
#> 0.07475574 0.13234870 0.09400592 0.11291828 0.10405537 0.02603028 0.04918370 
#>                                                                              
#> 0.08552612 0.08833234 0.01915353 0.03477198 0.05340467 0.03756540 0.03613504 
#>                                                                              
#> 0.08808790 0.22172547 0.11897425 0.07631712 0.09037474 0.06712969 0.04971851 
#>                                                                              
#> 0.03676111 0.06366187 0.13841810 0.13063248 0.33997868 0.20443603 0.06468976 
#>                                                                              
#> 0.11993673 0.05250916 0.03521725 0.07542635 0.10413650 0.03192010 0.15397348 
#>                                                                              
#> 0.21049505 0.16973029 0.23987857 0.25914901 0.06002921 0.16633648 0.20814014 
#>                                                                              
#> 0.12194184 0.08247795 0.19776074 0.10315635 0.05570390 0.13128613 0.17716021 
#>                                                                              
#> 0.10696676 0.07676522 0.09002202 0.06519728 0.09155301 0.09452581 0.14657108 
#>                                                                              
#> 0.11737169 0.18227092 0.09737292 0.17982693 0.18678737 0.10813173 0.67717971 
#>                                                                              
#> 0.10246596 0.17267298 0.06665918 0.11447156 0.05124337 0.03236083 0.06796315 
#>                                                                              
#> 0.15553428 0.11603210 0.10704803 0.13746764 0.15972801 0.19355867 0.18459635 
#>                                                                              
#> 0.10668581 0.19563087 0.09847415 0.09744832 0.11427227 0.06852619 0.20377124 
#>                                                                              
#> 0.17973953 0.09125902 0.27721465 0.25670186 0.25594182 0.10515076 0.02450578 
#>                                                                              
#> 0.09344214 0.14777719 0.09728304 0.10417569 0.11041962 0.10282902 0.12871173 
#>                                                                              
#> 0.14424434 0.11062501 0.13556193 0.01858263 0.09262313 0.10954154 0.15209636 
#>                                                                              
#> 0.08475466 0.12701635 0.21954857 0.15695997 0.15058896 0.32736114 0.10505332 
#>                                                                              
#> 0.13536291 0.09001638 0.12791963 0.15178384 0.03155993 0.10564448 0.12206961 
#>                                                                              
#> 0.22600293 0.12762947 0.10819951 0.19055531 0.06414310 0.12076538 0.11932721 
#>                                                                              
#> 0.18608729 0.13466779 0.15040051 0.11516560 0.12420582 0.04884077 0.11478162 
#>                                                                              
#> 0.07292927 0.02987266 0.36863752 0.09434199 0.02290595 0.02111931 0.11882595 
#>                                                                              
#> 0.04623423 0.18542488 0.05689957 0.11290174 0.12722030 0.17147118 0.13276173 
#>                                                                              
#> 0.14570295 0.12842614 0.17932753 0.06011109 0.11529408 0.23174743 0.12608000 
#>                                                                              
#> 0.14924159 0.11519447 0.12806805 0.06938059 0.17255991 0.31486473 0.31102055 
#>                                                                              
#> 0.14292312 0.16357861 0.49211091 0.12515729 0.12422296 0.10508410 0.09264464 
#>                                                                              
#> 0.12304617 0.13190184 0.10223068 0.10319048 0.17794249 0.17272831 0.15385239 
#>                                                                              
#> 0.18316487 0.09892352 0.07514949 0.15018219 0.25805377 0.08783213 0.10400534 
#>                                                                              
#> 0.11111514 0.12172227 0.14295431 0.07524779 0.16619530 0.06494414 0.09797033 
#>                                                                              
#> 0.08923066 0.08738498 0.06984741 0.07779651 0.01266356 0.16763451 0.13242033 
#>                                                                              
#> 0.16161789 0.27937978 0.20542537 0.18868457 0.14633141 0.14563294 0.21647395 
#>                                                                              
#> 0.09466929 0.14708699 0.11455586 0.14645674 0.04834457 0.12569541 0.11069156 
#>                                                                              
#> 0.14182013 0.13505428 0.04554126 0.15676357 0.07132172 0.13853893 0.10017152 
#>                                                                              
#> 0.17268281 0.15308938 0.01672601 0.52142533 0.23336973 0.07231616 0.19674034 
#>                                                        
#> 0.08450895 0.09523299 0.18914941 0.10349725 0.51007019 
#> 
#> $alpha
#> [1] 0.2583519
#> 
#> attr(,"class")
#> [1] "scree"