Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2          mx1
#>    91   60 0.309981634 0.407827954 0.450268010 0.429084494  0.375559814
#>    60   82 0.450268010 0.429084494 0.497546669 0.438830963  0.465795229
#>    82   93 0.497546669 0.438830963 0.538343037 0.548164878  0.493712254
#>    93   63 0.538343037 0.548164878 0.430263287 0.529152561  0.488384041
#>    63    3 0.430263287 0.529152561 0.281008308 0.572903617  0.345809476
#>     3   91 0.281008308 0.572903617 0.309981634 0.407827954  0.336505435
#>    79   20 0.444892962 0.177465824 0.478877381 0.261164321  0.407915523
#>    20   97 0.478877381 0.261164321 0.536448551 0.324129678  0.460623748
#>    97   60 0.536448551 0.324129678 0.450268010 0.429084494  0.459556463
#>    91    4 0.309981634 0.407827954 0.199095287 0.313031612  0.293378331
#>     4   11 0.199095287 0.313031612 0.321984618 0.201046182  0.293268627
#>    11   79 0.321984618 0.201046182 0.444892962 0.177465824  0.392121179
#>    63   34 0.430263287 0.529152561 0.390973792 0.667264654  0.365721962
#>    34   70 0.390973792 0.667264654 0.319835102 0.724333243  0.314632479
#>    70   19 0.319835102 0.724333243 0.224035631 0.646014822  0.302191430
#>    19  100 0.224035631 0.646014822 0.221426904 0.557992277  0.242213399
#>   100   91 0.221426904 0.557992277 0.309981634 0.407827954  0.271080715
#>    74   77 0.142844773 0.350997914 0.057739150 0.053089831  0.086001579
#>    77   57 0.057739150 0.053089831 0.173520516 0.036682731  0.134724185
#>    57   87 0.173520516 0.036682731 0.280123537 0.146732605  0.136196484
#>    87   11 0.280123537 0.146732605 0.321984618 0.201046182  0.232744807
#>    91    5 0.309981634 0.407827954 0.173898744 0.354319687  0.246241258
#>     5   74 0.173898744 0.354319687 0.142844773 0.350997914  0.162001377
#>    91   78 0.309981634 0.407827954 0.177214301 0.447954887  0.234877088
#>    78   74 0.177214301 0.447954887 0.142844773 0.350997914  0.153101072
#>    93   47 0.538343037 0.548164878 0.627079332 0.639021376  0.518306924
#>    47   10 0.627079332 0.639021376 0.545592080 0.834253813  0.547799058
#>    10   71 0.545592080 0.834253813 0.331272253 0.863183070  0.436792221
#>    71   55 0.331272253 0.863183070 0.366357684 0.715391438  0.386370202
#>    55   32 0.366357684 0.715391438 0.419927648 0.719343057  0.387730430
#>    32   34 0.419927648 0.719343057 0.390973792 0.667264654  0.463909176
#>    34   93 0.390973792 0.667264654 0.538343037 0.548164878  0.491559667
#>    84   80 0.224933823 0.848411010 0.042826771 0.835086560  0.132616758
#>    80   14 0.042826771 0.835086560 0.123813176 0.750052267  0.116081055
#>    14   19 0.123813176 0.750052267 0.224035631 0.646014822  0.178548842
#>    19   24 0.224035631 0.646014822 0.283783697 0.744163374  0.216410180
#>    24   84 0.283783697 0.744163374 0.224933823 0.848411010  0.234706067
#>    44   84 0.262778153 0.893541131 0.224933823 0.848411010  0.218849661
#>    84   86 0.224933823 0.848411010 0.277135246 0.851392605  0.250553734
#>    86   50 0.277135246 0.851392605 0.279700881 0.863022157  0.257784808
#>    50   44 0.279700881 0.863022157 0.262778153 0.893541131  0.286078072
#>    95   58 0.055626787 0.753443403 0.049604823 0.783651694  0.005238068
#>    58   80 0.049604823 0.783651694 0.042826771 0.835086560 -0.065594397
#>    80   51 0.042826771 0.835086560 0.005878550 0.938296282 -0.165103143
#>    51   52 0.005878550 0.938296282 0.003503652 0.949526123 -0.648436663
#>    52   96 0.003503652 0.949526123 0.000000000 0.440979106 -3.140428786
#>    96   89 0.000000000 0.440979106 0.043796171 0.605572889 -0.208589668
#>    89   38 0.043796171 0.605572889 0.053133963 0.727218483  0.055964487
#>    38   95 0.053133963 0.727218483 0.055626787 0.753443403  0.021233158
#>    64   40 0.740364392 0.721648369 0.748159563 0.746894321  0.636892040
#>    40   65 0.748159563 0.746894321 0.668042838 0.907382711  0.650173509
#>    65   13 0.668042838 0.907382711 0.563354266 1.000000000  0.580849531
#>    13   33 0.563354266 1.000000000 0.400916803 0.965181950  0.494796790
#>    33   71 0.400916803 0.965181950 0.331272253 0.863183070  0.440418536
#>    47   64 0.627079332 0.639021376 0.740364392 0.721648369  0.629512051
#>    87   72 0.280123537 0.146732605 0.328484835 0.178368542  0.292649284
#>    72   79 0.328484835 0.178368542 0.444892962 0.177465824  0.386917437
#>    56   98 0.881754934 0.498554541 0.947835331 0.630237217  0.848795697
#>    98   59 0.947835331 0.630237217 0.765688173 0.665709666  0.851305431
#>    59   64 0.765688173 0.665709666 0.740364392 0.721648369  0.718050137
#>    47   45 0.627079332 0.639021376 0.695262600 0.507342527  0.698160029
#>    45   37 0.695262600 0.507342527 0.730303294 0.503808771  0.720526975
#>    37   56 0.730303294 0.503808771 0.881754934 0.498554541  0.808244609
#>    31   94 0.184076151 0.988862043 0.252889107 0.986807056  0.296848095
#>    94   33 0.252889107 0.986807056 0.400916803 0.965181950  0.390393384
#>    65   39 0.668042838 0.907382711 0.695795270 0.958089939  0.624811298
#>    39   29 0.695795270 0.958089939 0.934772287 0.921904236  0.997020338
#>    38   88 0.053133963 0.727218483 0.137785123 0.636444770  0.104193437
#>    88   19 0.137785123 0.636444770 0.224035631 0.646014822  0.174541657
#>    99   62 0.603761070 0.165580829 0.616778699 0.145915237  0.643620021
#>    62   27 0.616778699 0.145915237 0.745957047 0.179812855  0.684585748
#>    27   92 0.745957047 0.179812855 0.709445287 0.294871990  0.720134383
#>    92   26 0.709445287 0.294871990 0.678209187 0.279580876  0.719136604
#>    26   21 0.678209187 0.279580876 0.646822443 0.267991983  0.672495966
#>    21   99 0.646822443 0.267991983 0.603761070 0.165580829  0.603167655
#>    28   81 1.000000000 0.010823122 0.452368168 0.016926881  0.649218946
#>    81   76 0.452368168 0.016926881 0.196375168 0.009691763  0.365619626
#>    76   57 0.196375168 0.009691763 0.173520516 0.036682731  0.162017494
#>    77   17 0.057739150 0.053089831 0.003858939 0.069454543 -0.027603045
#>    74   96 0.142844773 0.350997914 0.000000000 0.440979106 -0.017379255
#>    50   71 0.279700881 0.863022157 0.331272253 0.863183070  0.305480967
#>    94   44 0.252889107 0.986807056 0.262778153 0.893541131  0.310220763
#>    70   24 0.319835102 0.724333243 0.283783697 0.744163374  0.273657393
#>    88  100 0.137785123 0.636444770 0.221426904 0.557992277  0.185139164
#>    79   67 0.444892962 0.177465824 0.546307857 0.145676388  0.507850545
#>    67   99 0.546307857 0.145676388 0.603761070 0.165580829  0.550757222
#>    99   97 0.603761070 0.165580829 0.536448551 0.324129678  0.563158163
#>    26   48 0.678209187 0.279580876 0.658165021 0.339938934  0.651303566
#>    48   53 0.658165021 0.339938934 0.573466234 0.340424046  0.615641476
#>    53   97 0.573466234 0.340424046 0.536448551 0.324129678  0.580519333
#>    84   31 0.224933823 0.848411010 0.184076151 0.988862043  0.132834852
#>    43   30 0.464375209 0.091351307 0.506006637 0.039111042  0.509958295
#>    30   83 0.506006637 0.039111042 0.571145923 0.103431969  0.537315062
#>    83   67 0.571145923 0.103431969 0.546307857 0.145676388  0.565400538
#>    67   43 0.546307857 0.145676388 0.464375209 0.091351307  0.490783891
#>    70   71 0.319835102 0.724333243 0.331272253 0.863183070  0.334153252
#>    71   86 0.331272253 0.863183070 0.277135246 0.851392605  0.316675324
#>    66   56 0.748887197 0.390547380 0.881754934 0.498554541  0.816173505
#>    45   85 0.695262600 0.507342527 0.598535855 0.471013297  0.666416990
#>    85   48 0.598535855 0.471013297 0.658165021 0.339938934  0.666314916
#>    48   66 0.658165021 0.339938934 0.748887197 0.390547380  0.673958405
#>    92   48 0.709445287 0.294871990 0.658165021 0.339938934  0.680753463
#>    73   41 0.780497064 0.058527497 0.847697657 0.104720970  0.784185580
#>    41   36 0.847697657 0.104720970 0.804401325 0.278550152  0.834890189
#>    36   92 0.804401325 0.278550152 0.709445287 0.294871990  0.749633731
#>    62   73 0.616778699 0.145915237 0.780497064 0.058527497  0.697728266
#>    90   98 0.970550980 0.744261360 0.947835331 0.630237217  2.257411122
#>    56   75 0.881754934 0.498554541 0.848112187 0.352290654  1.265281427
#>    75   41 0.848112187 0.352290654 0.847697657 0.104720970  1.028923201
#>    41   49 0.847697657 0.104720970 0.897969154 0.043818825  0.997426306
#>    49   73 0.897969154 0.043818825 0.780497064 0.058527497  0.765018665
#>    73   61 0.780497064 0.058527497 0.547753135 0.026557103  0.752358924
#>    61   81 0.547753135 0.026557103 0.452368168 0.016926881  0.703781817
#>    39   42 0.695795270 0.958089939 0.855043303 0.901139068  0.841830336
#>    42   68 0.855043303 0.901139068 0.846331432 0.859052151  0.906078865
#>    68   90 0.846331432 0.859052151 0.970550980 0.744261360  0.934010539
#>    43   81 0.464375209 0.091351307 0.452368168 0.016926881  0.469100327
#>    61   83 0.547753135 0.026557103 0.571145923 0.103431969  0.538470840
#>    55   34 0.366357684 0.715391438 0.390973792 0.667264654  0.394467272
#>    55   70 0.366357684 0.715391438 0.319835102 0.724333243  0.333136011
#>    36   75 0.804401325 0.278550152 0.848112187 0.352290654  0.867320185
#>    75   66 0.848112187 0.352290654 0.748887197 0.390547380  0.799621529
#>    66   92 0.748887197 0.390547380 0.709445287 0.294871990  0.753932640
#>    89   88 0.043796171 0.605572889 0.137785123 0.636444770  0.076591928
#>    65   68 0.668042838 0.907382711 0.846331432 0.859052151  0.759152112
#>    59   90 0.765688173 0.665709666 0.970550980 0.744261360  0.859151590
#>    72   69 0.328484835 0.178368542 0.299601168 0.034003843  0.370797587
#>    69   81 0.299601168 0.034003843 0.452368168 0.016926881  0.380592641
#>    47   54 0.627079332 0.639021376 0.595702577 0.509025327  0.646452392
#>    54   85 0.595702577 0.509025327 0.598535855 0.471013297  0.645236184
#>    82   97 0.497546669 0.438830963 0.536448551 0.324129678  0.529129450
#>    54   93 0.595702577 0.509025327 0.538343037 0.548164878  0.542524043
#>    76   69 0.196375168 0.009691763 0.299601168 0.034003843  0.237243423
#>    69   87 0.299601168 0.034003843 0.280123537 0.146732605  0.237537887
#>    73   83 0.780497064 0.058527497 0.571145923 0.103431969  0.665472669
#>    99   83 0.603761070 0.165580829 0.571145923 0.103431969  0.582281060
#>    78   96 0.177214301 0.447954887 0.000000000 0.440979106  0.089392493
#>     1   60 0.429894327 0.503440897 0.450268010 0.429084494  0.375559814
#>     1   82 0.429894327 0.503440897 0.497546669 0.438830963  0.465795229
#>     1   93 0.429894327 0.503440897 0.538343037 0.548164878  0.493712254
#>     1   63 0.429894327 0.503440897 0.430263287 0.529152561  0.488384041
#>     1    3 0.429894327 0.503440897 0.281008308 0.572903617  0.345809476
#>     1   91 0.429894327 0.503440897 0.309981634 0.407827954  0.336505435
#>     2   20 0.387637138 0.312093279 0.478877381 0.261164321  0.407915523
#>     2   97 0.387637138 0.312093279 0.536448551 0.324129678  0.460623748
#>     2   60 0.387637138 0.312093279 0.450268010 0.429084494  0.459556463
#>     2   91 0.387637138 0.312093279 0.309981634 0.407827954  0.384590444
#>     2    4 0.387637138 0.312093279 0.199095287 0.313031612  0.293378331
#>     2   11 0.387637138 0.312093279 0.321984618 0.201046182  0.293268627
#>     2   79 0.387637138 0.312093279 0.444892962 0.177465824  0.392121179
#>     3   34 0.281008308 0.572903617 0.390973792 0.667264654  0.365721962
#>     3   70 0.281008308 0.572903617 0.319835102 0.724333243  0.314632479
#>     3   19 0.281008308 0.572903617 0.224035631 0.646014822  0.302191430
#>     3  100 0.281008308 0.572903617 0.221426904 0.557992277  0.242213399
#>     4   77 0.199095287 0.313031612 0.057739150 0.053089831  0.086001579
#>     4   57 0.199095287 0.313031612 0.173520516 0.036682731  0.134724185
#>     4   87 0.199095287 0.313031612 0.280123537 0.146732605  0.136196484
#>     4    5 0.199095287 0.313031612 0.173898744 0.354319687  0.246241258
#>     4   74 0.199095287 0.313031612 0.142844773 0.350997914  0.162001377
#>     5   78 0.173898744 0.354319687 0.177214301 0.447954887  0.234877088
#>     6   47 0.434306195 0.727772956 0.627079332 0.639021376  0.518306924
#>     6   10 0.434306195 0.727772956 0.545592080 0.834253813  0.547799058
#>     6   71 0.434306195 0.727772956 0.331272253 0.863183070  0.436792221
#>     6   55 0.434306195 0.727772956 0.366357684 0.715391438  0.386370202
#>     6   32 0.434306195 0.727772956 0.419927648 0.719343057  0.387730430
#>     6   34 0.434306195 0.727772956 0.390973792 0.667264654  0.463909176
#>     6   93 0.434306195 0.727772956 0.538343037 0.548164878  0.491559667
#>     7   80 0.171608553 0.774669751 0.042826771 0.835086560  0.132616758
#>     7   14 0.171608553 0.774669751 0.123813176 0.750052267  0.116081055
#>     7   19 0.171608553 0.774669751 0.224035631 0.646014822  0.178548842
#>     7   24 0.171608553 0.774669751 0.283783697 0.744163374  0.216410180
#>     7   84 0.171608553 0.774669751 0.224933823 0.848411010  0.234706067
#>     8   84 0.261951136 0.883312792 0.224933823 0.848411010  0.218849661
#>     8   86 0.261951136 0.883312792 0.277135246 0.851392605  0.250553734
#>     8   50 0.261951136 0.883312792 0.279700881 0.863022157  0.257784808
#>     8   44 0.261951136 0.883312792 0.262778153 0.893541131  0.286078072
#>     9   58 0.012087845 0.708862135 0.049604823 0.783651694  0.005238068
#>     9   80 0.012087845 0.708862135 0.042826771 0.835086560 -0.065594397
#>     9   51 0.012087845 0.708862135 0.005878550 0.938296282 -0.165103143
#>     9   52 0.012087845 0.708862135 0.003503652 0.949526123 -0.648436663
#>     9   96 0.012087845 0.708862135 0.000000000 0.440979106 -3.140428786
#>     9   89 0.012087845 0.708862135 0.043796171 0.605572889 -0.208589668
#>     9   38 0.012087845 0.708862135 0.053133963 0.727218483  0.055964487
#>     9   95 0.012087845 0.708862135 0.055626787 0.753443403  0.021233158
#>    10   40 0.545592080 0.834253813 0.748159563 0.746894321  0.636892040
#>    10   65 0.545592080 0.834253813 0.668042838 0.907382711  0.650173509
#>    10   13 0.545592080 0.834253813 0.563354266 1.000000000  0.580849531
#>    10   33 0.545592080 0.834253813 0.400916803 0.965181950  0.494796790
#>    10   64 0.545592080 0.834253813 0.740364392 0.721648369  0.629512051
#>    11   72 0.321984618 0.201046182 0.328484835 0.178368542  0.292649284
#>    12   98 0.759431852 0.651308247 0.947835331 0.630237217  0.848795697
#>    12   59 0.759431852 0.651308247 0.765688173 0.665709666  0.851305431
#>    12   64 0.759431852 0.651308247 0.740364392 0.721648369  0.718050137
#>    12   47 0.759431852 0.651308247 0.627079332 0.639021376  0.690904849
#>    12   45 0.759431852 0.651308247 0.695262600 0.507342527  0.698160029
#>    12   37 0.759431852 0.651308247 0.730303294 0.503808771  0.720526975
#>    12   56 0.759431852 0.651308247 0.881754934 0.498554541  0.808244609
#>    13   94 0.563354266 1.000000000 0.252889107 0.986807056  0.296848095
#>    13   39 0.563354266 1.000000000 0.695795270 0.958089939  0.624811298
#>    13   29 0.563354266 1.000000000 0.934772287 0.921904236  0.997020338
#>    14   58 0.123813176 0.750052267 0.049604823 0.783651694  0.109750403
#>    14   95 0.123813176 0.750052267 0.055626787 0.753443403  0.090935397
#>    14   38 0.123813176 0.750052267 0.053133963 0.727218483  0.088988574
#>    14   88 0.123813176 0.750052267 0.137785123 0.636444770  0.104193437
#>    15   62 0.662640931 0.214929927 0.616778699 0.145915237  0.643620021
#>    15   27 0.662640931 0.214929927 0.745957047 0.179812855  0.684585748
#>    15   92 0.662640931 0.214929927 0.709445287 0.294871990  0.720134383
#>    15   26 0.662640931 0.214929927 0.678209187 0.279580876  0.719136604
#>    15   21 0.662640931 0.214929927 0.646822443 0.267991983  0.672495966
#>    15   99 0.662640931 0.214929927 0.603761070 0.165580829  0.603167655
#>    16   81 0.127357921 0.000000000 0.452368168 0.016926881  0.649218946
#>    16   76 0.127357921 0.000000000 0.196375168 0.009691763  0.365619626
#>    16   57 0.127357921 0.000000000 0.173520516 0.036682731  0.162017494
#>    16   77 0.127357921 0.000000000 0.057739150 0.053089831  0.117705572
#>    16   17 0.127357921 0.000000000 0.003858939 0.069454543 -0.027603045
#>    17   74 0.003858939 0.069454543 0.142844773 0.350997914  0.075689466
#>    17   96 0.003858939 0.069454543 0.000000000 0.440979106 -0.017379255
#>    18   71 0.328053629 0.877491383 0.331272253 0.863183070  0.305480967
#>    18   33 0.328053629 0.877491383 0.400916803 0.965181950  0.405367754
#>    18   94 0.328053629 0.877491383 0.252889107 0.986807056  0.323850584
#>    18   44 0.328053629 0.877491383 0.262778153 0.893541131  0.310220763
#>    18   50 0.328053629 0.877491383 0.279700881 0.863022157  0.297172766
#>    20   67 0.478877381 0.261164321 0.546307857 0.145676388  0.507850545
#>    20   99 0.478877381 0.261164321 0.603761070 0.165580829  0.550757222
#>    21   48 0.646822443 0.267991983 0.658165021 0.339938934  0.651303566
#>    21   53 0.646822443 0.267991983 0.573466234 0.340424046  0.615641476
#>    21   97 0.646822443 0.267991983 0.536448551 0.324129678  0.580519333
#>    22   51 0.068964486 0.980496752 0.005878550 0.938296282  0.042439976
#>    22   80 0.068964486 0.980496752 0.042826771 0.835086560  0.074135030
#>    22   84 0.068964486 0.980496752 0.224933823 0.848411010  0.130022988
#>    22   31 0.068964486 0.980496752 0.184076151 0.988862043  0.132834852
#>    23   30 0.541880572 0.118127880 0.506006637 0.039111042  0.509958295
#>    23   83 0.541880572 0.118127880 0.571145923 0.103431969  0.537315062
#>    23   67 0.541880572 0.118127880 0.546307857 0.145676388  0.565400538
#>    23   43 0.541880572 0.118127880 0.464375209 0.091351307  0.490783891
#>    24   71 0.283783697 0.744163374 0.331272253 0.863183070  0.334153252
#>    24   86 0.283783697 0.744163374 0.277135246 0.851392605  0.316675324
#>    25   56 0.734768407 0.470051924 0.881754934 0.498554541  0.816173505
#>    25   37 0.734768407 0.470051924 0.730303294 0.503808771  0.805871227
#>    25   45 0.734768407 0.470051924 0.695262600 0.507342527  0.710610146
#>    25   85 0.734768407 0.470051924 0.598535855 0.471013297  0.666416990
#>    25   48 0.734768407 0.470051924 0.658165021 0.339938934  0.666314916
#>    25   66 0.734768407 0.470051924 0.748887197 0.390547380  0.673958405
#>    27   41 0.745957047 0.179812855 0.847697657 0.104720970  0.784185580
#>    27   36 0.745957047 0.179812855 0.804401325 0.278550152  0.834890189
#>    27   73 0.745957047 0.179812855 0.780497064 0.058527497  0.697728266
#>    28   98 1.000000000 0.010823122 0.947835331 0.630237217  2.257411122
#>    28   56 1.000000000 0.010823122 0.881754934 0.498554541  1.339422053
#>    28   75 1.000000000 0.010823122 0.848112187 0.352290654  1.265281427
#>    28   41 1.000000000 0.010823122 0.847697657 0.104720970  1.028923201
#>    28   49 1.000000000 0.010823122 0.897969154 0.043818825  0.997426306
#>    28   73 1.000000000 0.010823122 0.780497064 0.058527497  0.765018665
#>    28   61 1.000000000 0.010823122 0.547753135 0.026557103  0.752358924
#>    29   42 0.934772287 0.921904236 0.855043303 0.901139068  0.841830336
#>    29   68 0.934772287 0.921904236 0.846331432 0.859052151  0.906078865
#>    29   90 0.934772287 0.921904236 0.970550980 0.744261360  0.934010539
#>    30   81 0.506006637 0.039111042 0.452368168 0.016926881  0.469100327
#>    30   61 0.506006637 0.039111042 0.547753135 0.026557103  0.505963227
#>    31   44 0.184076151 0.988862043 0.262778153 0.893541131  0.191523368
#>    35   75 0.811310612 0.322363464 0.848112187 0.352290654  0.867320185
#>    35   66 0.811310612 0.322363464 0.748887197 0.390547380  0.799621529
#>    35   92 0.811310612 0.322363464 0.709445287 0.294871990  0.753932640
#>    35   36 0.811310612 0.322363464 0.804401325 0.278550152  0.760567854
#>    39   68 0.695795270 0.958089939 0.846331432 0.859052151  0.759152112
#>    40   59 0.748159563 0.746894321 0.765688173 0.665709666  0.802779957
#>    40   90 0.748159563 0.746894321 0.970550980 0.744261360  0.859151590
#>    40   68 0.748159563 0.746894321 0.846331432 0.859052151  0.859390785
#>    43   79 0.464375209 0.091351307 0.444892962 0.177465824  0.489563082
#>    43   72 0.464375209 0.091351307 0.328484835 0.178368542  0.386231493
#>    43   69 0.464375209 0.091351307 0.299601168 0.034003843  0.370797587
#>    45   54 0.695262600 0.507342527 0.595702577 0.509025327  0.646452392
#>    46   82 0.559291568 0.439648551 0.497546669 0.438830963  0.527749106
#>    46   97 0.559291568 0.439648551 0.536448551 0.324129678  0.529129450
#>    46   53 0.559291568 0.439648551 0.573466234 0.340424046  0.531712814
#>    46   48 0.559291568 0.439648551 0.658165021 0.339938934  0.616141887
#>    46   85 0.559291568 0.439648551 0.598535855 0.471013297  0.621316144
#>    46   54 0.559291568 0.439648551 0.595702577 0.509025327  0.553772331
#>    46   93 0.559291568 0.439648551 0.538343037 0.548164878  0.542524043
#>    57   69 0.173520516 0.036682731 0.299601168 0.034003843  0.237243423
#>    62   83 0.616778699 0.145915237 0.571145923 0.103431969  0.582281060
#>    78   88 0.177214301 0.447954887 0.137785123 0.636444770  0.120800075
#>    78   89 0.177214301 0.447954887 0.043796171 0.605572889  0.119301081
#>    78  100 0.177214301 0.447954887 0.221426904 0.557992277  0.259038473
#>    13   31 0.563354266 1.000000000 0.184076151 0.988862043  0.296848095
#>    16   28 0.127357921 0.000000000 1.000000000 0.010823122  0.649218946
#>    22   52 0.068964486 0.980496752 0.003503652 0.949526123  0.042439976
#>    28   90 1.000000000 0.010823122 0.970550980 0.744261360  2.257411122
#>           my1          mx2          my2 bp1 bp2
#>   0.448583819  0.384590444  0.388984543   0   0
#>   0.473308358  0.486834453  0.371250239   0   0
#>   0.502539964  0.527749106  0.489839605   0   0
#>   0.515460054  0.470815362  0.615333123   0   0
#>   0.517505987  0.365721962  0.585436635   0   0
#>   0.497563754  0.271080715  0.486080707   0   0
#>   0.241228574  0.507850545  0.200651578   0   0
#>   0.335656409  0.563158163  0.241906028   0   0
#>   0.348851725  0.486834453  0.371250239   0   0
#>   0.314997529  0.246241258  0.370135289   0   0
#>   0.292954328  0.232744807  0.226537369   0   0
#>   0.234511385  0.386917437  0.207387819   0   0
#>   0.585436635  0.470815362  0.615333123   0   0
#>   0.644974765  0.333136011  0.668040292   0   0
#>   0.648164670  0.273657393  0.683067641   0   0
#>   0.601426157  0.185139164  0.603117668   0   0
#>   0.486080707  0.259038473  0.478979169   0   0
#>   0.206126313  0.075689466  0.209072251   0   0
#>   0.179630998  0.117705572  0.059534320   0   0
#>   0.179494744  0.237537887  0.081327438   0   0
#>   0.226537369  0.292649284  0.180367217   0   0
#>   0.370135289  0.234877088  0.399036787   0   0
#>   0.318726893  0.153101072  0.401932419   0   0
#>   0.399036787  0.259038473  0.478979169   0   0
#>   0.401932419  0.089392493  0.424515981   0   0
#>   0.656494467  0.599911610  0.576794086   0   0
#>   0.720552944  0.629512051  0.754658789   0   0
#>   0.836569048  0.440418536  0.863434284   0   0
#>   0.798202790  0.356804668  0.791183992   0   0
#>   0.790738000  0.394467272  0.699410287   0   0
#>   0.660802904  0.394467272  0.699410287   0   0
#>   0.641001287  0.470815362  0.615333123   0   0
#>   0.859017738  0.130022988  0.894467145   0   0
#>   0.823770969  0.109750403  0.817741675   0   0
#>   0.702488408  0.174541657  0.698628163   0   0
#>   0.717916960  0.273657393  0.683067641   0   0
#>   0.785192854  0.254105028  0.796143956   0   0
#>   0.891945385  0.191523368  0.914860131   0   0
#>   0.858319606  0.254105028  0.796143956   0   0
#>   0.861759355  0.305523620  0.851227531   0   0
#>   0.886509602  0.297172766  0.892661595   0   0
#>   0.759102889  0.090935397  0.776186484   0   0
#>   0.794634855  0.109750403  0.817741675   0   0
#>   0.818867820  0.074135030  0.904513094   0   0
#>   0.805787115  0.042439976  0.951894371   0   0
#>   0.716900778 -5.106721978  0.730447624   0   1
#>   0.584605660  0.086178227  0.506171925   0   0
#>   0.665820014  0.076591928  0.664236604   0   0
#>   0.743481770  0.088988574  0.737041242   0   0
#>   0.767423868  0.802779957  0.716202799   0   0
#>   0.798220680  0.747288215  0.846700898   0   0
#>   0.914300285  0.624811298  0.963991812   0   0
#>   0.923522124  0.390393384  1.410597767   0   0
#>   0.863434284  0.405367754  0.887366861   0   0
#>   0.754658789  0.690904849  0.670486697   0   0
#>   0.180367217  0.345261258  0.099940222   0   0
#>   0.207387819  0.386231493  0.118933403   0   0
#>   0.597515414  1.339422053  0.351311541   0   0
#>   0.619955826  0.867950503  0.705426457   0   0
#>   0.677845110  0.802779957  0.716202799   0   0
#>   0.592334877  0.646452392  0.565560678   0   0
#>   0.582365350  0.710610146  0.484030173   0   0
#>   0.565042656  0.805871227  0.496630624   0   0
#>   3.611967997  0.216929808  0.935837035   0   0
#>   1.410597767  0.323850584  0.955100484   0   0
#>   0.963991812  0.759152112  0.890466115   0   0
#>   2.140218659  0.841830336  1.115315422   0   0
#>   0.689976432  0.076591928  0.664236604   0   0
#>   0.698628163  0.185139164  0.603117668   0   0
#>   0.177824138  0.582281060  0.137220846   0   0
#>   0.150601243  0.697728266  0.100517229   0   0
#>   0.234941252  0.749633731  0.244302291   0   0
#>   0.235525430  0.680753463  0.313933038   0   0
#>   0.246756717  0.651303566  0.304153090   0   0
#>   0.226089047  0.564348686  0.242411469   0   0
#>  -6.891469810  0.703781817 -1.996065845   0   0
#>  -1.446126165  0.334513459 -0.345527574   0   0
#>   0.003770934  0.237243423  0.067468624   0   0
#>  -0.131014549  0.075689466  0.209072251   0   0
#>   0.255016269  0.089392493  0.424515981   0   0
#>   0.864897544  0.305523620  0.851227531   0   0
#>   0.945728735  0.216929808  0.935837035   0   0
#>   0.683067641  0.334153252  0.793049804   0   0
#>   0.603117668  0.120800075  0.534522830   0   0
#>   0.200651578  0.489563082  0.142310780   0   0
#>   0.225703699  0.581218592  0.137778420   0   0
#>   0.241906028  0.564348686  0.242411469   0   0
#>   0.304153090  0.680753463  0.313933038   0   0
#>   0.309775289  0.616141887  0.397145165   0   0
#>   0.274205030  0.531712814  0.385084093   0   0
#>   0.897787447  0.191523368  0.914860131   0   0
#>   0.084968843  0.469100327  0.052408221   0   0
#>   0.072548770  0.538470840  0.071378287   0   0
#>   0.128478023  0.581218592  0.137778420   0   0
#>   0.140469566  0.489563082  0.142310780   0   0
#>   0.793049804  0.356804668  0.791183992   0   0
#>   0.800023453  0.305523620  0.851227531   0   0
#>   0.443502311  0.823036614  0.435059483   0   0
#>   0.437211776  0.645236184  0.493605775   0   0
#>   0.422747153  0.621316144  0.402276031   0   0
#>   0.418247086  0.712192840  0.349706922   0   0
#>   0.313933038  0.712192840  0.349706922   0   0
#>   0.125138809  0.838585373  0.045999960   0   0
#>   0.193837549  0.973236250  0.228295958   0   0
#>   0.244302291  0.760567854  0.307914066   0   0
#>   0.100517229  0.668110230  0.045028708   0   0
#>   0.428621044  0.867950503  0.705426457   0   0
#>   0.333336970  0.823036614  0.435059483   0   0
#>   0.228202716  0.973236250  0.228295958   0   0
#>   0.177114741  0.838585373  0.045999960   0   0
#>  -0.541546970  0.838585373  0.045999960   0   0
#>  -0.599798450  0.665472669  0.032732017   0   0
#>  -1.996065845  0.505963227 -0.036721565   0   0
#>   1.115315422  0.764118591  0.898015100   0   0
#>   0.868629728  0.764118591  0.898015100   0   0
#>   0.829326319  0.859390785  0.748577400   0   0
#>   0.052408221  0.380592641  0.066687344   0   0
#>   0.071378287  0.665472669  0.032732017   0   0
#>   0.699410287  0.333136011  0.668040292   0   0
#>   0.668040292  0.356804668  0.791183992   0   0
#>   0.291079399  0.973236250  0.228295958   0   0
#>   0.374328682  0.823036614  0.435059483   0   0
#>   0.332499807  0.712192840  0.349706922   0   0
#>   0.664236604  0.119301081  0.534209264   0   0
#>   0.890466115  0.747288215  0.846700898   0   0
#>   0.728374019  0.867950503  0.705426457   0   0
#>   0.094831058  0.345261258  0.099940222   0   0
#>   0.066687344  0.334513459 -0.345527574   0   0
#>   0.565560678  0.599911610  0.576794086   0   0
#>   0.493605775  0.553772331  0.486788394   0   0
#>   0.385594934  0.486834453  0.371250239   0   0
#>   0.492691833  0.599911610  0.576794086   0   0
#>   0.067468624  0.334513459 -0.345527574   0   0
#>   0.081327438  0.345261258  0.099940222   0   0
#>   0.032732017  0.668110230  0.045028708   0   0
#>   0.137220846  0.581218592  0.137778420   0   0
#>   0.424515981  0.086178227  0.506171925   0   0
#>   0.448583819  0.465795229  0.473308358   0   0
#>   0.473308358  0.493712254  0.502539964   0   0
#>   0.502539964  0.488384041  0.515460054   0   0
#>   0.515460054  0.345809476  0.517505987   0   0
#>   0.517505987  0.336505435  0.497563754   0   0
#>   0.497563754  0.375559814  0.448583819   0   0
#>   0.241228574  0.460623748  0.335656409   0   0
#>   0.335656409  0.459556463  0.348851725   0   0
#>   0.348851725  0.384590444  0.388984543   0   0
#>   0.388984543  0.293378331  0.314997529   0   0
#>   0.314997529  0.293268627  0.292954328   0   0
#>   0.292954328  0.392121179  0.234511385   0   0
#>   0.234511385  0.407915523  0.241228574   0   0
#>   0.585436635  0.314632479  0.644974765   0   0
#>   0.644974765  0.302191430  0.648164670   0   0
#>   0.648164670  0.242213399  0.601426157   0   0
#>   0.601426157  0.271080715  0.486080707   0   0
#>   0.206126313  0.134724185  0.179630998   0   0
#>   0.179630998  0.136196484  0.179494744   0   0
#>   0.179494744  0.232744807  0.226537369   0   0
#>   0.370135289  0.162001377  0.318726893   0   0
#>   0.318726893  0.086001579  0.206126313   0   0
#>   0.399036787  0.153101072  0.401932419   0   0
#>   0.656494467  0.547799058  0.720552944   0   0
#>   0.720552944  0.436792221  0.836569048   0   0
#>   0.836569048  0.386370202  0.798202790   0   0
#>   0.798202790  0.387730430  0.790738000   0   0
#>   0.790738000  0.463909176  0.660802904   0   0
#>   0.660802904  0.491559667  0.641001287   0   0
#>   0.641001287  0.518306924  0.656494467   0   0
#>   0.859017738  0.116081055  0.823770969   0   0
#>   0.823770969  0.178548842  0.702488408   0   0
#>   0.702488408  0.216410180  0.717916960   0   0
#>   0.717916960  0.234706067  0.785192854   0   0
#>   0.785192854  0.132616758  0.859017738   0   0
#>   0.891945385  0.250553734  0.858319606   0   0
#>   0.858319606  0.257784808  0.861759355   0   0
#>   0.861759355  0.286078072  0.886509602   0   0
#>   0.886509602  0.218849661  0.891945385   0   0
#>   0.759102889 -0.065594397  0.794634855   0   0
#>   0.794634855 -0.165103143  0.818867820   0   0
#>   0.818867820 -0.648436663  0.805787115   0   0
#>   0.805787115 -3.140428786  0.716900778   0   0
#>   0.716900778 -0.208589668  0.584605660   0   0
#>   0.584605660  0.055964487  0.665820014   0   0
#>   0.665820014  0.021233158  0.743481770   0   0
#>   0.743481770  0.005238068  0.759102889   0   0
#>   0.767423868  0.650173509  0.798220680   0   0
#>   0.798220680  0.580849531  0.914300285   0   0
#>   0.914300285  0.494796790  0.923522124   0   0
#>   0.923522124  0.440418536  0.863434284   0   0
#>   0.754658789  0.636892040  0.767423868   0   0
#>   0.180367217  0.386917437  0.207387819   0   0
#>   0.597515414  0.851305431  0.619955826   0   0
#>   0.619955826  0.718050137  0.677845110   0   0
#>   0.677845110  0.690904849  0.670486697   0   0
#>   0.670486697  0.698160029  0.592334877   0   0
#>   0.592334877  0.720526975  0.582365350   0   0
#>   0.582365350  0.808244609  0.565042656   0   0
#>   0.565042656  0.848795697  0.597515414   0   0
#>   3.611967997  0.390393384  1.410597767   0   0
#>   0.963991812  0.997020338  2.140218659   0   0
#>   2.140218659  1.539162678  4.718609919   0   1
#>   0.817741675  0.090935397  0.776186484   0   0
#>   0.776186484  0.088988574  0.737041242   0   0
#>   0.737041242  0.104193437  0.689976432   0   0
#>   0.689976432  0.174541657  0.698628163   0   0
#>   0.177824138  0.684585748  0.150601243   0   0
#>   0.150601243  0.720134383  0.234941252   0   0
#>   0.234941252  0.719136604  0.235525430   0   0
#>   0.235525430  0.672495966  0.246756717   0   0
#>   0.246756717  0.603167655  0.226089047   0   0
#>   0.226089047  0.643620021  0.177824138   0   0
#>  -6.891469810  0.365619626 -1.446126165   0   0
#>  -1.446126165  0.162017494  0.003770934   0   0
#>   0.003770934  0.117705572  0.059534320   0   0
#>   0.059534320 -0.027603045 -0.131014549   0   0
#>  -0.131014549 -3.487195460 -6.282608469   0   1
#>   0.209072251 -0.017379255  0.255016269   0   0
#>   0.255016269 -2.708700973  0.227062140   0   1
#>   0.864897544  0.405367754  0.887366861   0   0
#>   0.887366861  0.323850584  0.955100484   0   0
#>   0.955100484  0.310220763  0.945728735   0   0
#>   0.945728735  0.297172766  0.892661595   0   0
#>   0.892661595  0.305480967  0.864897544   0   0
#>   0.200651578  0.550757222  0.225703699   0   0
#>   0.225703699  0.563158163  0.241906028   0   0
#>   0.304153090  0.615641476  0.309775289   0   0
#>   0.309775289  0.580519333  0.274205030   0   0
#>   0.274205030  0.564348686  0.242411469   0   0
#>   0.951894371  0.074135030  0.904513094   0   0
#>   0.904513094  0.130022988  0.894467145   0   0
#>   0.894467145  0.132834852  0.897787447   0   0
#>   0.897787447 -0.495158244  9.539367314   0   1
#>   0.084968843  0.537315062  0.072548770   0   0
#>   0.072548770  0.565400538  0.128478023   0   0
#>   0.128478023  0.490783891  0.140469566   0   0
#>   0.140469566  0.509958295  0.084968843   0   0
#>   0.793049804  0.316675324  0.800023453   0   0
#>   0.800023453  0.254105028  0.796143956   0   0
#>   0.443502311  0.805871227  0.496630624   0   0
#>   0.496630624  0.710610146  0.484030173   0   0
#>   0.484030173  0.666416990  0.437211776   0   0
#>   0.437211776  0.666314916  0.422747153   0   0
#>   0.422747153  0.673958405  0.418247086   0   0
#>   0.418247086  0.816173505  0.443502311   0   0
#>   0.125138809  0.834890189  0.193837549   0   0
#>   0.193837549  0.749633731  0.244302291   0   0
#>   0.100517229  0.784185580  0.125138809   0   0
#>   0.428621044  1.339422053  0.351311541   0   0
#>   0.351311541  1.265281427  0.333336970   0   0
#>   0.333336970  1.028923201  0.228202716   0   0
#>   0.228202716  0.997426306  0.177114741   0   0
#>   0.177114741  0.765018665 -0.541546970   0   0
#>  -0.541546970  0.752358924 -0.599798450   0   0
#>  -0.599798450  0.703781817 -1.996065845   0   0
#>   1.115315422  0.906078865  0.868629728   0   0
#>   0.868629728  0.934010539  0.829326319   0   0
#>   0.829326319  6.343831814  1.918907513   0   1
#>   0.052408221  0.505963227 -0.036721565   0   0
#>  -0.036721565  0.538470840  0.071378287   0   0
#>   0.914860131  0.216929808  0.935837035   0   0
#>   0.291079399  0.799621529  0.374328682   0   0
#>   0.374328682  0.753932640  0.332499807   0   0
#>   0.332499807  0.760567854  0.307914066   0   0
#>   0.307914066  0.867320185  0.291079399   0   0
#>   0.890466115  0.764118591  0.898015100   0   0
#>   0.716202799  0.859151590  0.728374019   0   0
#>   0.728374019  0.859390785  0.748577400   0   0
#>   0.748577400  0.747288215  0.846700898   0   0
#>   0.142310780  0.386231493  0.118933403   0   0
#>   0.118933403  0.370797587  0.094831058   0   0
#>   0.094831058  0.380592641  0.066687344   0   0
#>   0.565560678  0.645236184  0.493605775   0   0
#>   0.489839605  0.529129450  0.385594934   0   0
#>   0.385594934  0.531712814  0.385084093   0   0
#>   0.385084093  0.616141887  0.397145165   0   0
#>   0.397145165  0.621316144  0.402276031   0   0
#>   0.402276031  0.553772331  0.486788394   0   0
#>   0.486788394  0.542524043  0.492691833   0   0
#>   0.492691833  0.527749106  0.489839605   0   0
#>   0.067468624  0.237537887  0.081327438   0   0
#>   0.137220846  0.668110230  0.045028708   0   0
#>   0.534522830  0.119301081  0.534209264   0   0
#>   0.534209264  0.086178227  0.506171925   0   0
#>   0.478979169  0.120800075  0.534522830   0   0
#>   3.611967997  0.219488247  6.246283905   0   1
#>  -6.891469810  0.663429580 -8.037238758   0   1
#>   0.951894371 -5.863138635 13.434175416   0   1
#>   0.428621044  3.618657863  0.483277838   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.429894327 0.503440897
#>   [2,] 0.387637138 0.312093279
#>   [3,] 0.281008308 0.572903617
#>   [4,] 0.199095287 0.313031612
#>   [5,] 0.173898744 0.354319687
#>   [6,] 0.434306195 0.727772956
#>   [7,] 0.171608553 0.774669751
#>   [8,] 0.261951136 0.883312792
#>   [9,] 0.012087845 0.708862135
#>  [10,] 0.545592080 0.834253813
#>  [11,] 0.321984618 0.201046182
#>  [12,] 0.759431852 0.651308247
#>  [13,] 0.563354266 1.000000000
#>  [14,] 0.123813176 0.750052267
#>  [15,] 0.662640931 0.214929927
#>  [16,] 0.127357921 0.000000000
#>  [17,] 0.003858939 0.069454543
#>  [18,] 0.328053629 0.877491383
#>  [19,] 0.224035631 0.646014822
#>  [20,] 0.478877381 0.261164321
#>  [21,] 0.646822443 0.267991983
#>  [22,] 0.068964486 0.980496752
#>  [23,] 0.541880572 0.118127880
#>  [24,] 0.283783697 0.744163374
#>  [25,] 0.734768407 0.470051924
#>  [26,] 0.678209187 0.279580876
#>  [27,] 0.745957047 0.179812855
#>  [28,] 1.000000000 0.010823122
#>  [29,] 0.934772287 0.921904236
#>  [30,] 0.506006637 0.039111042
#>  [31,] 0.184076151 0.988862043
#>  [32,] 0.419927648 0.719343057
#>  [33,] 0.400916803 0.965181950
#>  [34,] 0.390973792 0.667264654
#>  [35,] 0.811310612 0.322363464
#>  [36,] 0.804401325 0.278550152
#>  [37,] 0.730303294 0.503808771
#>  [38,] 0.053133963 0.727218483
#>  [39,] 0.695795270 0.958089939
#>  [40,] 0.748159563 0.746894321
#>  [41,] 0.847697657 0.104720970
#>  [42,] 0.855043303 0.901139068
#>  [43,] 0.464375209 0.091351307
#>  [44,] 0.262778153 0.893541131
#>  [45,] 0.695262600 0.507342527
#>  [46,] 0.559291568 0.439648551
#>  [47,] 0.627079332 0.639021376
#>  [48,] 0.658165021 0.339938934
#>  [49,] 0.897969154 0.043818825
#>  [50,] 0.279700881 0.863022157
#>  [51,] 0.005878550 0.938296282
#>  [52,] 0.003503652 0.949526123
#>  [53,] 0.573466234 0.340424046
#>  [54,] 0.595702577 0.509025327
#>  [55,] 0.366357684 0.715391438
#>  [56,] 0.881754934 0.498554541
#>  [57,] 0.173520516 0.036682731
#>  [58,] 0.049604823 0.783651694
#>  [59,] 0.765688173 0.665709666
#>  [60,] 0.450268010 0.429084494
#>  [61,] 0.547753135 0.026557103
#>  [62,] 0.616778699 0.145915237
#>  [63,] 0.430263287 0.529152561
#>  [64,] 0.740364392 0.721648369
#>  [65,] 0.668042838 0.907382711
#>  [66,] 0.748887197 0.390547380
#>  [67,] 0.546307857 0.145676388
#>  [68,] 0.846331432 0.859052151
#>  [69,] 0.299601168 0.034003843
#>  [70,] 0.319835102 0.724333243
#>  [71,] 0.331272253 0.863183070
#>  [72,] 0.328484835 0.178368542
#>  [73,] 0.780497064 0.058527497
#>  [74,] 0.142844773 0.350997914
#>  [75,] 0.848112187 0.352290654
#>  [76,] 0.196375168 0.009691763
#>  [77,] 0.057739150 0.053089831
#>  [78,] 0.177214301 0.447954887
#>  [79,] 0.444892962 0.177465824
#>  [80,] 0.042826771 0.835086560
#>  [81,] 0.452368168 0.016926881
#>  [82,] 0.497546669 0.438830963
#>  [83,] 0.571145923 0.103431969
#>  [84,] 0.224933823 0.848411010
#>  [85,] 0.598535855 0.471013297
#>  [86,] 0.277135246 0.851392605
#>  [87,] 0.280123537 0.146732605
#>  [88,] 0.137785123 0.636444770
#>  [89,] 0.043796171 0.605572889
#>  [90,] 0.970550980 0.744261360
#>  [91,] 0.309981634 0.407827954
#>  [92,] 0.709445287 0.294871990
#>  [93,] 0.538343037 0.548164878
#>  [94,] 0.252889107 0.986807056
#>  [95,] 0.055626787 0.753443403
#>  [96,] 0.000000000 0.440979106
#>  [97,] 0.536448551 0.324129678
#>  [98,] 0.947835331 0.630237217
#>  [99,] 0.603761070 0.165580829
#> [100,] 0.221426904 0.557992277
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.4298943,0.5034409) [6]: 3 60 63 82 91 93 
#> 2: (0.3876371,0.3120933) [7]: 4 11 20 60 79 91 97 
#> 3: (0.2810083,0.5729036) [7]: 1 19 34 63 70 91 100 
#> 4: (0.1990953,0.3130316) [8]: 2 5 11 57 74 77 87 91 
#> 5: (0.1738987,0.3543197) [4]: 4 74 78 91 
#> 6: (0.4343062,0.727773) [7]: 10 32 34 47 55 71 93 
#> 7: (0.1716086,0.7746698) [5]: 14 19 24 80 84 
#> 8: (0.2619511,0.8833128) [4]: 44 50 84 86 
#> 9: (0.01208785,0.7088621) [8]: 38 51 52 58 80 89 95 96 
#> 10: (0.5455921,0.8342538) [8]: 6 13 33 40 47 64 65 71 
#> 11: (0.3219846,0.2010462) [5]: 2 4 72 79 87 
#> 12: (0.7594319,0.6513082) [7]: 37 45 47 56 59 64 98 
#> 13: (0.5633543,1) [7]: 10 29 31 33 39 65 94 
#> 14: (0.1238132,0.7500523) [7]: 7 19 38 58 80 88 95 
#> 15: (0.6626409,0.2149299) [6]: 21 26 27 62 92 99 
#> 16: (0.1273579,0) [6]: 17 28 57 76 77 81 
#> 17: (0.003858939,0.06945454) [4]: 16 74 77 96 
#> 18: (0.3280536,0.8774914) [5]: 33 44 50 71 94 
#> 19: (0.2240356,0.6460148) [7]: 3 7 14 24 70 88 100 
#> 20: (0.4788774,0.2611643) [5]: 2 67 79 97 99 
#> 21: (0.6468224,0.267992) [6]: 15 26 48 53 97 99 
#> 22: (0.06896449,0.9804968) [5]: 31 51 52 80 84 
#> 23: (0.5418806,0.1181279) [4]: 30 43 67 83 
#> 24: (0.2837837,0.7441634) [6]: 7 19 70 71 84 86 
#> 25: (0.7347684,0.4700519) [6]: 37 45 48 56 66 85 
#> 26: (0.6782092,0.2795809) [4]: 15 21 48 92 
#> 27: (0.745957,0.1798129) [6]: 15 36 41 62 73 92 
#> 28: (1,0.01082312) [10]: 16 41 49 56 61 73 75 81 90 98 
#> 29: (0.9347723,0.9219042) [5]: 13 39 42 68 90 
#> 30: (0.5060066,0.03911104) [5]: 23 43 61 81 83 
#> 31: (0.1840762,0.988862) [5]: 13 22 44 84 94 
#> 32: (0.4199276,0.7193431) [3]: 6 34 55 
#> 33: (0.4009168,0.9651819) [5]: 10 13 18 71 94 
#> 34: (0.3909738,0.6672647) [7]: 3 6 32 55 63 70 93 
#> 35: (0.8113106,0.3223635) [4]: 36 66 75 92 
#> 36: (0.8044013,0.2785502) [5]: 27 35 41 75 92 
#> 37: (0.7303033,0.5038088) [4]: 12 25 45 56 
#> 38: (0.05313396,0.7272185) [5]: 9 14 88 89 95 
#> 39: (0.6957953,0.9580899) [5]: 13 29 42 65 68 
#> 40: (0.7481596,0.7468943) [6]: 10 59 64 65 68 90 
#> 41: (0.8476977,0.104721) [6]: 27 28 36 49 73 75 
#> 42: (0.8550433,0.9011391) [3]: 29 39 68 
#> 43: (0.4643752,0.09135131) [7]: 23 30 67 69 72 79 81 
#> 44: (0.2627782,0.8935411) [6]: 8 18 31 50 84 94 
#> 45: (0.6952626,0.5073425) [6]: 12 25 37 47 54 85 
#> 46: (0.5592916,0.4396486) [7]: 48 53 54 82 85 93 97 
#> 47: (0.6270793,0.6390214) [7]: 6 10 12 45 54 64 93 
#> 48: (0.658165,0.3399389) [8]: 21 25 26 46 53 66 85 92 
#> 49: (0.8979692,0.04381883) [3]: 28 41 73 
#> 50: (0.2797009,0.8630222) [5]: 8 18 44 71 86 
#> 51: (0.00587855,0.9382963) [4]: 9 22 52 80 
#> 52: (0.003503652,0.9495261) [4]: 9 22 51 96 
#> 53: (0.5734662,0.340424) [4]: 21 46 48 97 
#> 54: (0.5957026,0.5090253) [5]: 45 46 47 85 93 
#> 55: (0.3663577,0.7153914) [5]: 6 32 34 70 71 
#> 56: (0.8817549,0.4985545) [7]: 12 25 28 37 66 75 98 
#> 57: (0.1735205,0.03668273) [6]: 4 16 69 76 77 87 
#> 58: (0.04960482,0.7836517) [4]: 9 14 80 95 
#> 59: (0.7656882,0.6657097) [5]: 12 40 64 90 98 
#> 60: (0.450268,0.4290845) [5]: 1 2 82 91 97 
#> 61: (0.5477531,0.0265571) [5]: 28 30 73 81 83 
#> 62: (0.6167787,0.1459152) [5]: 15 27 73 83 99 
#> 63: (0.4302633,0.5291526) [4]: 1 3 34 93 
#> 64: (0.7403644,0.7216484) [5]: 10 12 40 47 59 
#> 65: (0.6680428,0.9073827) [5]: 10 13 39 40 68 
#> 66: (0.7488872,0.3905474) [6]: 25 35 48 56 75 92 
#> 67: (0.5463079,0.1456764) [6]: 20 23 43 79 83 99 
#> 68: (0.8463314,0.8590522) [6]: 29 39 40 42 65 90 
#> 69: (0.2996012,0.03400384) [6]: 43 57 72 76 81 87 
#> 70: (0.3198351,0.7243332) [6]: 3 19 24 34 55 71 
#> 71: (0.3312723,0.8631831) [9]: 6 10 18 24 33 50 55 70 86 
#> 72: (0.3284848,0.1783685) [5]: 11 43 69 79 87 
#> 73: (0.7804971,0.0585275) [7]: 27 28 41 49 61 62 83 
#> 74: (0.1428448,0.3509979) [6]: 4 5 17 77 78 96 
#> 75: (0.8481122,0.3522907) [6]: 28 35 36 41 56 66 
#> 76: (0.1963752,0.009691763) [4]: 16 57 69 81 
#> 77: (0.05773915,0.05308983) [5]: 4 16 17 57 74 
#> 78: (0.1772143,0.4479549) [7]: 5 74 88 89 91 96 100 
#> 79: (0.444893,0.1774658) [6]: 2 11 20 43 67 72 
#> 80: (0.04282677,0.8350866) [7]: 7 9 14 22 51 58 84 
#> 81: (0.4523682,0.01692688) [7]: 16 28 30 43 61 69 76 
#> 82: (0.4975467,0.438831) [5]: 1 46 60 93 97 
#> 83: (0.5711459,0.103432) [7]: 23 30 61 62 67 73 99 
#> 84: (0.2249338,0.848411) [8]: 7 8 22 24 31 44 80 86 
#> 85: (0.5985359,0.4710133) [5]: 25 45 46 48 54 
#> 86: (0.2771352,0.8513926) [5]: 8 24 50 71 84 
#> 87: (0.2801235,0.1467326) [5]: 4 11 57 69 72 
#> 88: (0.1377851,0.6364448) [6]: 14 19 38 78 89 100 
#> 89: (0.04379617,0.6055729) [5]: 9 38 78 88 96 
#> 90: (0.970551,0.7442614) [6]: 28 29 40 59 68 98 
#> 91: (0.3099816,0.407828) [8]: 1 2 3 4 5 60 78 100 
#> 92: (0.7094453,0.294872) [7]: 15 26 27 35 36 48 66 
#> 93: (0.538343,0.5481649) [8]: 1 6 34 46 47 54 63 82 
#> 94: (0.2528891,0.9868071) [5]: 13 18 31 33 44 
#> 95: (0.05562679,0.7534434) [4]: 9 14 38 58 
#> 96: (0,0.4409791) [6]: 9 17 52 74 78 89 
#> 97: (0.5364486,0.3241297) [8]: 2 20 21 46 53 60 82 99 
#> 98: (0.9478353,0.6302372) [5]: 12 28 56 59 90 
#> 99: (0.6037611,0.1655808) [7]: 15 20 21 62 67 83 97 
#> 100: (0.2214269,0.5579923) [5]: 3 19 78 88 91 
#> number of nodes: 100 
#> number of arcs: 287 
#> number of boundary nodes: 10 
#> boundary nodes:  13 16 17 22 28 29 31 52 90 96 
#> number of triangles: 188 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.14188766 0.04827282 0.11669725 0.10973924 0.15553522 0.16759901 0.09033482 
#>                                                                              
#> 0.08531750 0.13580353 0.14588396 0.16626041 0.12514989 0.14359183 0.09120053 
#>                                                                              
#> 0.12373889 0.08806119 0.17433090 0.30982607 0.11693809 0.15321612 0.06857343 
#>                                                                              
#> 0.14622478 0.03123113 0.13869872 0.10286845 0.12700013 0.21155585 0.21626347 
#>                                                                              
#> 0.15189916 0.05371551 0.05958595 0.18947942 0.18259386 0.11742925 0.14445875 
#>                                                                              
#> 0.11490418 0.11971164 0.05889755 0.05228650 0.01190920 0.03489680 0.03080268 
#>                                                                              
#> 0.05187955 0.10962398 0.01147822 0.50855909 0.17032093 0.12200346 0.02634313 
#>                                                                              
#> 0.02642201 0.17937450 0.13977718 0.16612714 0.12350763 0.14021671 0.05778968 
#>                                                                              
#> 0.11641163 0.14733277 0.18556908 0.06140385 0.14828445 0.03521843 0.15154275 
#>                                                                              
#> 0.06884363 0.14959894 0.05780502 0.24170110 0.12411964 0.08677981 0.02358377 
#>                                                                              
#> 0.13355184 0.12071335 0.03477804 0.03345789 0.11109602 0.54766585 0.25609522 
#>                                                                              
#> 0.03536732 0.05631058 0.16882311 0.05157162 0.09378873 0.04114533 0.11467668 
#>                                                                              
#> 0.10628052 0.06080344 0.17224608 0.06359924 0.08470018 0.04044521 0.14627318 
#>                                                                              
#> 0.06679986 0.09154402 0.04900531 0.09830653 0.13932007 0.05540605 0.17122903 
#>                                                                              
#> 0.10332413 0.14400044 0.10388324 0.06826928 0.08154604 0.17914005 0.09634859 
#>                                                                              
#> 0.18558103 0.11626481 0.15008317 0.24757003 0.07897021 0.11838934 0.23492944 
#>                                                                              
#> 0.09586988 0.16912521 0.04297913 0.16913729 0.07538676 0.08035526 0.05405682 
#>                                                                              
#> 0.04737411 0.08572223 0.10634461 0.10348645 0.09892925 0.18472321 0.21940633 
#>                                                                              
#> 0.14722579 0.15371851 0.13372910 0.03811747 0.12111871 0.06944078 0.10605039 
#>                                                                              
#> 0.11439909 0.21411285 0.07018710 0.17735154 0.07709709 0.09354829 0.11730881 
#>                                                                              
#> 0.02571431 0.16429278 0.15336521 0.10449182 0.14929739 0.13270106 0.12327005 
#>                                                                              
#> 0.18854419 0.12900276 0.14629689 0.14490139 0.15632803 0.09268837 0.06141899 
#>                                                                              
#> 0.29589067 0.27752977 0.18498902 0.04836911 0.06786428 0.09369388 0.21222235 
#>                                                                              
#> 0.15402182 0.17015256 0.06906737 0.01666751 0.07442413 0.20756379 0.14224956 
#>                                                                              
#> 0.05376261 0.13892692 0.11624931 0.09100197 0.05087648 0.03534764 0.02695855 
#>                                                                              
#> 0.01026172 0.08367199 0.12991338 0.22951815 0.24081703 0.26815561 0.10804669 
#>                                                                              
#> 0.04496376 0.06231476 0.22060205 0.14262547 0.16669521 0.19512333 0.22498053 
#>                                                                              
#> 0.02359085 0.18957811 0.01570167 0.07287867 0.13292162 0.15761923 0.15034816 
#>                                                                              
#> 0.19569525 0.31074534 0.13891390 0.37953958 0.08146043 0.06827066 0.07427606 
#>                                                                              
#> 0.11446344 0.08286357 0.09041451 0.09263574 0.06649899 0.05536972 0.07682559 
#>                                                                              
#> 0.32545073 0.06969441 0.05896277 0.08755172 0.14168956 0.31398046 0.37154460 
#>                                                                              
#> 0.01466586 0.11401174 0.13266357 0.06721966 0.05047125 0.13373231 0.15726455 
#>                                                                              
#> 0.07283555 0.10308995 0.12382987 0.07589937 0.14774066 0.20438463 0.11541522 
#>                                                                              
#> 0.08677903 0.03274799 0.02790199 0.08200040 0.12814387 0.10743514 0.14972454 
#>                                                                              
#> 0.03405087 0.05432585 0.13623594 0.15098831 0.08074845 0.12645135 0.11473791 
#>                                                                              
#> 0.12610770 0.62160677 0.50186037 0.37372447 0.17892124 0.10723344 0.22462690 
#>                                                                              
#> 0.45252048 0.08238873 0.10849963 0.18121012 0.05804500 0.04359325 0.12361263 
#>                                                                              
#> 0.04743409 0.09244311 0.10550984 0.04435476 0.18019328 0.08305541 0.22240700 
#>                                                                              
#> 0.14905400 0.08829081 0.16136354 0.17446838 0.09957424 0.06175031 0.11775574 
#>                                                                              
#> 0.10023185 0.14042068 0.05023805 0.07835112 0.11051984 0.12610911 0.06234724 
#>                                                                              
#> 0.19256972 0.20650383 0.11858744 0.37944162 0.87270919 0.07241754 0.73402922 
#> 
#> $alpha
#> [1] 0.2820352
#> 
#> attr(,"class")
#> [1] "scree"