Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, global_rng = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

global_rng

an optional vector c(min, max) that specifies the global range of the data set if you want all variables one the same scale

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    84   43 0.502032840 0.88534824 0.413135669 0.95391046  0.456111275
#>    43   40 0.413135669 0.95391046 0.363591254 0.96698511  0.382163133
#>    40   20 0.363591254 0.96698511 0.327216356 0.95885013  0.354997656
#>    20   31 0.327216356 0.95885013 0.274637595 0.91777343  0.336163195
#>    31    2 0.274637595 0.91777343 0.266972559 0.79342073  0.332647487
#>     2   29 0.266972559 0.79342073 0.302817263 0.71384765  0.379680659
#>    29   84 0.302817263 0.71384765 0.502032840 0.88534824  0.425093939
#>    31   98 0.274637595 0.91777343 0.192616751 0.87218709  0.254973577
#>    98   33 0.192616751 0.87218709 0.172034642 0.73944645  0.198542996
#>    33   29 0.172034642 0.73944645 0.302817263 0.71384765  0.238628952
#>    68   15 0.464586561 0.02326433 0.674084279 0.02363838  0.569892147
#>    15   12 0.674084279 0.02363838 0.600150232 0.14836312  0.640013462
#>    12   68 0.600150232 0.14836312 0.464586561 0.02326433  0.514538909
#>    60   87 0.692918611 0.54702676 0.853550284 0.50640290  0.771201111
#>    87   72 0.853550284 0.50640290 0.852700151 0.62967405  0.798176800
#>    72   36 0.852700151 0.62967405 0.841616454 0.66052754  0.779430293
#>    36   96 0.841616454 0.66052754 0.716976327 0.74131417  0.753726917
#>    96   38 0.716976327 0.74131417 0.648254753 0.60042297  0.701767364
#>    38   60 0.648254753 0.60042297 0.692918611 0.54702676  0.679929687
#>    59   37 0.271741561 0.51905751 0.174314380 0.50461114  0.234056059
#>    37   94 0.174314380 0.50461114 0.146443590 0.31642126  0.177057689
#>    94   18 0.146443590 0.31642126 0.183683104 0.29162835  0.195265236
#>    18   80 0.183683104 0.29162835 0.230387120 0.29011435  0.208416025
#>    80    8 0.230387120 0.29011435 0.298652542 0.17358113  0.345101470
#>     8   27 0.298652542 0.17358113 0.449181841 0.42172103  0.387104951
#>    27   16 0.449181841 0.42172103 0.384144478 0.50328800  0.344586612
#>    16   59 0.384144478 0.50328800 0.271741561 0.51905751  0.316402019
#>    55   88 0.332381530 0.10705598 0.578492400 0.19742487  0.444968475
#>    88   76 0.578492400 0.19742487 0.696318575 0.21519042  0.628939687
#>    76   30 0.696318575 0.21519042 0.586347667 0.43164117  0.652826317
#>    30   10 0.586347667 0.43164117 0.524303772 0.38577352  0.605608866
#>    10   27 0.524303772 0.38577352 0.449181841 0.42172103  0.441436205
#>     8   55 0.298652542 0.17358113 0.332381530 0.10705598  0.431230777
#>    66   51 0.992357830 0.45484500 0.994336143 0.91407418  2.531309820
#>    51   64 0.994336143 0.91407418 0.910721047 0.91526440  0.952317674
#>    64   23 0.910721047 0.91526440 0.911430666 0.81228123  0.933601674
#>    23   83 0.911430666 0.81228123 0.948315664 0.73488240  0.985156943
#>    83   61 0.948315664 0.73488240 0.967531225 0.61139937  1.260376947
#>    61   66 0.967531225 0.61139937 0.992357830 0.45484500  1.856251061
#>    80   48 0.230387120 0.29011435 0.171211789 0.26351246  0.230082360
#>    48   97 0.171211789 0.26351246 0.123612094 0.12301143  0.203645528
#>    97  100 0.123612094 0.12301143 0.259022478 0.09694256  0.203668271
#>   100   55 0.259022478 0.09694256 0.332381530 0.10705598  0.292058897
#>    77   32 0.950634189 0.04195001 1.000000000 0.11217075  0.958303647
#>    32   62 1.000000000 0.11217075 0.995894455 0.38477589  1.246526191
#>    62   44 0.995894455 0.38477589 0.826594588 0.23588068  0.965723070
#>    44   77 0.826594588 0.23588068 0.950634189 0.04195001  0.887283827
#>    30   47 0.586347667 0.43164117 0.547655673 0.43705233  0.562128945
#>    47   27 0.547655673 0.43705233 0.449181841 0.42172103  0.498564348
#>    68   78 0.464586561 0.02326433 0.411034805 0.03532491  0.375933323
#>    78   95 0.411034805 0.03532491 0.311056991 0.05712693  0.338300917
#>    95  100 0.311056991 0.05712693 0.259022478 0.09694256  0.260630512
#>   100   50 0.259022478 0.09694256 0.168431644 0.01941880  0.215346451
#>    50   54 0.168431644 0.01941880 0.017911473 0.01058693  0.111651923
#>    15   89 0.674084279 0.02363838 0.740012236 0.14308601  0.668530390
#>    89   76 0.740012236 0.14308601 0.696318575 0.21519042  0.670246335
#>    88   68 0.578492400 0.19742487 0.464586561 0.02326433  0.491755171
#>    69   79 0.364349889 1.00000000 0.790228455 0.89926084  0.665833249
#>    79   41 0.790228455 0.89926084 0.819137974 0.85246686  0.845868677
#>    41   64 0.819137974 0.85246686 0.910721047 0.91526440  0.860750211
#>    95   91 0.311056991 0.05712693 0.365509437 0.06554892  0.337162434
#>    91   55 0.365509437 0.06554892 0.332381530 0.10705598  0.338401463
#>    77   89 0.950634189 0.04195001 0.740012236 0.14308601  0.812878624
#>    47   73 0.547655673 0.43705233 0.604146575 0.47734634  0.502392418
#>    73   38 0.604146575 0.47734634 0.648254753 0.60042297  0.504994385
#>    38   22 0.648254753 0.60042297 0.345646778 0.53751200  0.480105207
#>    22   59 0.345646778 0.53751200 0.271741561 0.51905751  0.322555739
#>    33   65 0.172034642 0.73944645 0.056485565 0.71483733  0.126722862
#>    65   63 0.056485565 0.71483733 0.002550851 0.72154719  0.011572850
#>    63   37 0.002550851 0.72154719 0.174314380 0.50461114 -0.042633469
#>    37   39 0.174314380 0.50461114 0.202762036 0.60270766  0.192546287
#>    39   33 0.202762036 0.60270766 0.172034642 0.73944645  0.162128782
#>    94   56 0.146443590 0.31642126 0.156313864 0.29825278  0.176272185
#>    56   48 0.156313864 0.29825278 0.171211789 0.26351246  0.166924026
#>    72   92 0.852700151 0.62967405 0.953849294 0.64590044  0.905454625
#>    92   83 0.953849294 0.64590044 0.948315664 0.73488240  0.941472124
#>    83   25 0.948315664 0.73488240 0.886478556 0.76472510  0.904094515
#>    25   34 0.886478556 0.76472510 0.823973755 0.75288900  0.863147165
#>    34   36 0.823973755 0.75288900 0.841616454 0.66052754  0.857397315
#>    69   67 0.364349889 1.00000000 0.177984329 0.97692593  0.263041685
#>    67   31 0.177984329 0.97692593 0.274637595 0.91777343  0.255985706
#>    40   69 0.363591254 0.96698511 0.364349889 1.00000000  0.340682692
#>    28   49 0.000000000 0.27666490 0.002968025 0.25044319  0.036623621
#>    49   35 0.002968025 0.25044319 0.019942336 0.13858695  0.056580762
#>    35   97 0.019942336 0.13858695 0.123612094 0.12301143  0.080894456
#>    48   70 0.171211789 0.26351246 0.125738959 0.28419285  0.117264856
#>    70   94 0.125738959 0.28419285 0.146443590 0.31642126  0.087139937
#>    94   28 0.146443590 0.31642126 0.000000000 0.27666490  0.045136458
#>    38   29 0.648254753 0.60042297 0.302817263 0.71384765  0.477425008
#>    29   85 0.302817263 0.71384765 0.293581543 0.53984284  0.323501103
#>    85   59 0.293581543 0.53984284 0.271741561 0.51905751  0.317559509
#>    41   93 0.819137974 0.85246686 0.837288444 0.77983262  0.861883519
#>    93   25 0.837288444 0.77983262 0.886478556 0.76472510  0.871319621
#>    79   75 0.790228455 0.89926084 0.625281126 0.80146003  0.713482491
#>    75   96 0.625281126 0.80146003 0.716976327 0.74131417  0.713391029
#>    96   34 0.716976327 0.74131417 0.823973755 0.75288900  0.769803263
#>    34   93 0.823973755 0.75288900 0.837288444 0.77983262  0.812063023
#>    42   54 0.069138552 0.06744611 0.017911473 0.01058693  0.090165380
#>    50   97 0.168431644 0.01941880 0.123612094 0.12301143  0.152621288
#>    97   42 0.123612094 0.12301143 0.069138552 0.06744611  0.077734028
#>    54   49 0.017911473 0.01058693 0.002968025 0.25044319 -2.623832900
#>    39   85 0.202762036 0.60270766 0.293581543 0.53984284  0.287046816
#>    38   75 0.648254753 0.60042297 0.625281126 0.80146003  0.484194886
#>    75   84 0.625281126 0.80146003 0.502032840 0.88534824  0.475815377
#>    76   46 0.696318575 0.21519042 0.688625178 0.44817400  0.655179762
#>    46   74 0.688625178 0.44817400 0.618320495 0.47930784  0.638939004
#>    74   73 0.618320495 0.47930784 0.604146575 0.47734634  0.615630268
#>    67   98 0.177984329 0.97692593 0.192616751 0.87218709  0.214891386
#>    51   62 0.994336143 0.91407418 0.995894455 0.38477589 34.090557979
#>    98   71 0.192616751 0.87218709 0.097078538 0.78794874  0.163681506
#>    71   65 0.097078538 0.78794874 0.056485565 0.71483733  0.113428601
#>    59   39 0.271741561 0.51905751 0.202762036 0.60270766  0.218097252
#>    74   60 0.618320495 0.47930784 0.692918611 0.54702676  0.630832654
#>    86   69 0.382415275 0.98363092 0.364349889 1.00000000  0.365803029
#>    43   86 0.413135669 0.95391046 0.382415275 0.98363092  0.387853331
#>    84   53 0.502032840 0.88534824 0.530385443 0.88367918  0.523291603
#>    53   79 0.530385443 0.88367918 0.790228455 0.89926084  0.641659549
#>    79   86 0.790228455 0.89926084 0.382415275 0.98363092  0.644799763
#>    62   90 0.995894455 0.38477589 0.856522406 0.35149867  0.950731326
#>    90   52 0.856522406 0.35149867 0.780161371 0.36183439  0.810892192
#>    52   76 0.780161371 0.36183439 0.696318575 0.21519042  0.752774466
#>    63   81 0.002550851 0.72154719 0.057128382 0.83409397 -0.171718270
#>    81   82 0.057128382 0.83409397 0.075023163 0.88350354  0.006788834
#>    82   58 0.075023163 0.88350354 0.103495036 0.91970713  0.076611942
#>    58   67 0.103495036 0.91970713 0.177984329 0.97692593  0.094999086
#>    52   87 0.780161371 0.36183439 0.853550284 0.50640290  0.779663414
#>    56   70 0.156313864 0.29825278 0.125738959 0.28419285  0.148743254
#>    66   62 0.992357830 0.45484500 0.995894455 0.38477589  5.825078722
#>    90   87 0.856522406 0.35149867 0.853550284 0.50640290  0.828055738
#>    91   88 0.365509437 0.06554892 0.578492400 0.19742487  0.450025754
#>    71   81 0.097078538 0.78794874 0.057128382 0.83409397  0.041378890
#>    82   98 0.075023163 0.88350354 0.192616751 0.87218709  0.132810429
#>    87   66 0.853550284 0.50640290 0.992357830 0.45484500  0.940107187
#>    99   90 0.960373145 0.42130023 0.856522406 0.35149867  0.929230977
#>    66   99 0.992357830 0.45484500 0.960373145 0.42130023  0.995448418
#>    87   99 0.853550284 0.50640290 0.960373145 0.42130023  0.925251540
#>    88   78 0.578492400 0.19742487 0.411034805 0.03532491  0.464470283
#>     1   43 0.400948620 0.90705319 0.413135669 0.95391046  0.456111275
#>     1   40 0.400948620 0.90705319 0.363591254 0.96698511  0.382163133
#>     1   20 0.400948620 0.90705319 0.327216356 0.95885013  0.354997656
#>     1   31 0.400948620 0.90705319 0.274637595 0.91777343  0.336163195
#>     1    2 0.400948620 0.90705319 0.266972559 0.79342073  0.332647487
#>     1   29 0.400948620 0.90705319 0.302817263 0.71384765  0.379680659
#>     1   84 0.400948620 0.90705319 0.502032840 0.88534824  0.425093939
#>     2   98 0.266972559 0.79342073 0.192616751 0.87218709  0.254973577
#>     2   33 0.266972559 0.79342073 0.172034642 0.73944645  0.198542996
#>     3   15 0.585139214 0.04022252 0.674084279 0.02363838  0.569892147
#>     3   12 0.585139214 0.04022252 0.600150232 0.14836312  0.640013462
#>     3   68 0.585139214 0.04022252 0.464586561 0.02326433  0.514538909
#>     4   87 0.716805733 0.58170752 0.853550284 0.50640290  0.771201111
#>     4   72 0.716805733 0.58170752 0.852700151 0.62967405  0.798176800
#>     4   36 0.716805733 0.58170752 0.841616454 0.66052754  0.779430293
#>     4   96 0.716805733 0.58170752 0.716976327 0.74131417  0.753726917
#>     4   38 0.716805733 0.58170752 0.648254753 0.60042297  0.701767364
#>     4   60 0.716805733 0.58170752 0.692918611 0.54702676  0.679929687
#>     5   37 0.254173502 0.34986118 0.174314380 0.50461114  0.234056059
#>     5   94 0.254173502 0.34986118 0.146443590 0.31642126  0.177057689
#>     5   18 0.254173502 0.34986118 0.183683104 0.29162835  0.195265236
#>     5   80 0.254173502 0.34986118 0.230387120 0.29011435  0.208416025
#>     5    8 0.254173502 0.34986118 0.298652542 0.17358113  0.345101470
#>     5   27 0.254173502 0.34986118 0.449181841 0.42172103  0.387104951
#>     5   16 0.254173502 0.34986118 0.384144478 0.50328800  0.344586612
#>     5   59 0.254173502 0.34986118 0.271741561 0.51905751  0.316402019
#>     6   88 0.546868194 0.26863166 0.578492400 0.19742487  0.444968475
#>     6   76 0.546868194 0.26863166 0.696318575 0.21519042  0.628939687
#>     6   30 0.546868194 0.26863166 0.586347667 0.43164117  0.652826317
#>     6   10 0.546868194 0.26863166 0.524303772 0.38577352  0.605608866
#>     6   27 0.546868194 0.26863166 0.449181841 0.42172103  0.441436205
#>     6    8 0.546868194 0.26863166 0.298652542 0.17358113  0.399358803
#>     6   55 0.546868194 0.26863166 0.332381530 0.10705598  0.431230777
#>     7   51 0.988784907 0.87459371 0.994336143 0.91407418  2.531309820
#>     7   64 0.988784907 0.87459371 0.910721047 0.91526440  0.952317674
#>     7   23 0.988784907 0.87459371 0.911430666 0.81228123  0.933601674
#>     7   83 0.988784907 0.87459371 0.948315664 0.73488240  0.985156943
#>     7   61 0.988784907 0.87459371 0.967531225 0.61139937  1.260376947
#>     7   66 0.988784907 0.87459371 0.992357830 0.45484500  1.856251061
#>     8   48 0.298652542 0.17358113 0.171211789 0.26351246  0.230082360
#>     8   97 0.298652542 0.17358113 0.123612094 0.12301143  0.203645528
#>     8  100 0.298652542 0.17358113 0.259022478 0.09694256  0.203668271
#>     9   32 0.999606101 0.11286642 1.000000000 0.11217075  0.958303647
#>     9   62 0.999606101 0.11286642 0.995894455 0.38477589  1.246526191
#>     9   44 0.999606101 0.11286642 0.826594588 0.23588068  0.965723070
#>     9   77 0.999606101 0.11286642 0.950634189 0.04195001  0.887283827
#>    10   47 0.524303772 0.38577352 0.547655673 0.43705233  0.562128945
#>    11   68 0.235145072 0.00000000 0.464586561 0.02326433  0.573290763
#>    11   78 0.235145072 0.00000000 0.411034805 0.03532491  0.375933323
#>    11   95 0.235145072 0.00000000 0.311056991 0.05712693  0.338300917
#>    11  100 0.235145072 0.00000000 0.259022478 0.09694256  0.260630512
#>    11   50 0.235145072 0.00000000 0.168431644 0.01941880  0.215346451
#>    11   54 0.235145072 0.00000000 0.017911473 0.01058693  0.111651923
#>    12   89 0.600150232 0.14836312 0.740012236 0.14308601  0.668530390
#>    12   76 0.600150232 0.14836312 0.696318575 0.21519042  0.670246335
#>    12   88 0.600150232 0.14836312 0.578492400 0.19742487  0.639128393
#>    13   79 0.885894451 0.94001195 0.790228455 0.89926084  0.665833249
#>    13   41 0.885894451 0.94001195 0.819137974 0.85246686  0.845868677
#>    13   64 0.885894451 0.94001195 0.910721047 0.91526440  0.860750211
#>    13   51 0.885894451 0.94001195 0.994336143 0.91407418  0.953501368
#>    14   95 0.309009156 0.07307424 0.311056991 0.05712693  0.272189548
#>    14   91 0.309009156 0.07307424 0.365509437 0.06554892  0.337162434
#>    14   55 0.309009156 0.07307424 0.332381530 0.10705598  0.338401463
#>    14  100 0.309009156 0.07307424 0.259022478 0.09694256  0.294901486
#>    16   73 0.384144478 0.50328800 0.604146575 0.47734634  0.502392418
#>    16   38 0.384144478 0.50328800 0.648254753 0.60042297  0.504994385
#>    16   22 0.384144478 0.50328800 0.345646778 0.53751200  0.480105207
#>    16   47 0.384144478 0.50328800 0.547655673 0.43705233  0.484803787
#>    17   65 0.158489001 0.59077979 0.056485565 0.71483733  0.126722862
#>    17   63 0.158489001 0.59077979 0.002550851 0.72154719  0.011572850
#>    17   37 0.158489001 0.59077979 0.174314380 0.50461114 -0.042633469
#>    17   39 0.158489001 0.59077979 0.202762036 0.60270766  0.192546287
#>    17   33 0.158489001 0.59077979 0.172034642 0.73944645  0.162128782
#>    18   56 0.183683104 0.29162835 0.156313864 0.29825278  0.176272185
#>    18   48 0.183683104 0.29162835 0.171211789 0.26351246  0.166924026
#>    19   92 0.897793164 0.67668015 0.953849294 0.64590044  0.905454625
#>    19   83 0.897793164 0.67668015 0.948315664 0.73488240  0.941472124
#>    19   25 0.897793164 0.67668015 0.886478556 0.76472510  0.904094515
#>    19   34 0.897793164 0.67668015 0.823973755 0.75288900  0.863147165
#>    19   36 0.897793164 0.67668015 0.841616454 0.66052754  0.857397315
#>    19   72 0.897793164 0.67668015 0.852700151 0.62967405  0.873719167
#>    20   67 0.327216356 0.95885013 0.177984329 0.97692593  0.263041685
#>    20   69 0.327216356 0.95885013 0.364349889 1.00000000  0.340682692
#>    21   49 0.074115065 0.27190189 0.002968025 0.25044319  0.036623621
#>    21   35 0.074115065 0.27190189 0.019942336 0.13858695  0.056580762
#>    21   97 0.074115065 0.27190189 0.123612094 0.12301143  0.080894456
#>    21   48 0.074115065 0.27190189 0.171211789 0.26351246  0.117117916
#>    21   70 0.074115065 0.27190189 0.125738959 0.28419285  0.117264856
#>    21   94 0.074115065 0.27190189 0.146443590 0.31642126  0.087139937
#>    21   28 0.074115065 0.27190189 0.000000000 0.27666490  0.045136458
#>    22   29 0.345646778 0.53751200 0.302817263 0.71384765  0.477425008
#>    22   85 0.345646778 0.53751200 0.293581543 0.53984284  0.323501103
#>    23   41 0.911430666 0.81228123 0.819137974 0.85246686  0.878859160
#>    23   93 0.911430666 0.81228123 0.837288444 0.77983262  0.861883519
#>    23   25 0.911430666 0.81228123 0.886478556 0.76472510  0.871319621
#>    24   79 0.800152301 0.79818414 0.790228455 0.89926084  0.754244059
#>    24   75 0.800152301 0.79818414 0.625281126 0.80146003  0.713482491
#>    24   96 0.800152301 0.79818414 0.716976327 0.74131417  0.713391029
#>    24   34 0.800152301 0.79818414 0.823973755 0.75288900  0.769803263
#>    24   93 0.800152301 0.79818414 0.837288444 0.77983262  0.812063023
#>    24   41 0.800152301 0.79818414 0.819137974 0.85246686  0.832684997
#>    26   54 0.095854042 0.07029665 0.017911473 0.01058693  0.090165380
#>    26   50 0.095854042 0.07029665 0.168431644 0.01941880  0.094565108
#>    26   97 0.095854042 0.07029665 0.123612094 0.12301143  0.152621288
#>    26   42 0.095854042 0.07029665 0.069138552 0.06744611  0.077734028
#>    28   37 0.000000000 0.27666490 0.174314380 0.50461114  0.037333327
#>    28   63 0.000000000 0.27666490 0.002550851 0.72154719 -0.055106282
#>    29   39 0.302817263 0.71384765 0.202762036 0.60270766  0.228350195
#>    29   75 0.302817263 0.71384765 0.625281126 0.80146003  0.484194886
#>    30   46 0.586347667 0.43164117 0.688625178 0.44817400  0.655179762
#>    30   74 0.586347667 0.43164117 0.618320495 0.47930784  0.638939004
#>    30   73 0.586347667 0.43164117 0.604146575 0.47734634  0.615630268
#>    33   71 0.172034642 0.73944645 0.097078538 0.78794874  0.163681506
#>    35   42 0.019942336 0.13858695 0.069138552 0.06744611  0.003777047
#>    35   54 0.019942336 0.13858695 0.017911473 0.01058693 -0.703325559
#>    38   74 0.648254753 0.60042297 0.618320495 0.47930784  0.601634755
#>    40   86 0.363591254 0.96698511 0.382415275 0.98363092  0.387853331
#>    43   53 0.413135669 0.95391046 0.530385443 0.88367918  0.523291603
#>    43   79 0.413135669 0.95391046 0.790228455 0.89926084  0.641659549
#>    44   90 0.826594588 0.23588068 0.856522406 0.35149867  0.950731326
#>    44   52 0.826594588 0.23588068 0.780161371 0.36183439  0.810892192
#>    44   76 0.826594588 0.23588068 0.696318575 0.21519042  0.752774466
#>    44   89 0.826594588 0.23588068 0.740012236 0.14308601  0.764377834
#>    45   81 0.054624106 0.92903723 0.057128382 0.83409397 -0.171718270
#>    45   82 0.054624106 0.92903723 0.075023163 0.88350354  0.006788834
#>    45   58 0.054624106 0.92903723 0.103495036 0.91970713  0.076611942
#>    45   67 0.054624106 0.92903723 0.177984329 0.97692593  0.094999086
#>    46   52 0.688625178 0.44817400 0.780161371 0.36183439  0.664358194
#>    46   87 0.688625178 0.44817400 0.853550284 0.50640290  0.779663414
#>    46   60 0.688625178 0.44817400 0.692918611 0.54702676  0.765055442
#>    53   75 0.530385443 0.88367918 0.625281126 0.80146003  0.509067773
#>    57   71 0.047055641 0.72412913 0.097078538 0.78794874  0.080938862
#>    57   81 0.047055641 0.72412913 0.057128382 0.83409397  0.041378890
#>    57   63 0.047055641 0.72412913 0.002550851 0.72154719  0.021375320
#>    57   65 0.047055641 0.72412913 0.056485565 0.71483733  0.026486538
#>    58   98 0.103495036 0.91970713 0.192616751 0.87218709  0.132810429
#>    61   87 0.967531225 0.61139937 0.853550284 0.50640290  0.901814347
#>    61   92 0.967531225 0.61139937 0.953849294 0.64590044  1.147123430
#>    61   72 0.967531225 0.61139937 0.852700151 0.62967405  0.908087010
#>    62   99 0.995894455 0.38477589 0.960373145 0.42130023  0.995448418
#>    71   82 0.097078538 0.78794874 0.075023163 0.88350354  0.130759754
#>    78   91 0.411034805 0.03532491 0.365509437 0.06554892  0.457562088
#>    11   15 0.235145072 0.00000000 0.674084279 0.02363838  0.573290763
#>    13   69 0.885894451 0.94001195 0.364349889 1.00000000  0.665833249
#>    15   77 0.674084279 0.02363838 0.950634189 0.04195001  0.812878624
#>    28   54 0.000000000 0.27666490 0.017911473 0.01058693 -2.623832900
#>    32   51 1.000000000 0.11217075 0.994336143 0.91407418 34.090557979
#>    45   63 0.054624106 0.92903723 0.002550851 0.72154719 -0.171718270
#>           my1          mx2          my2 bp1 bp2
#>   0.917719496  0.523291603  1.004824917   0   0
#>   0.936952578  0.387853331  0.958514722   0   0
#>   0.920019519  0.340682692  0.984027677   0   0
#>   0.893208907  0.255985706  0.995837256   0   0
#>   0.851785147  0.254973577  0.856572926   0   0
#>   0.796331638  0.238628952  0.732793100   0   0
#>   0.773265701  0.475815377  0.714347530   0   0
#>   0.856572926  0.214891386  0.928690464   0   0
#>   0.803302196  0.163681506  0.808707647   0   0
#>   0.732793100  0.228350195  0.680279596   0   0
#>  -0.288361333  0.573290763 -2.191867715   0   0
#>   0.087717558  0.668530390  0.104621758   0   0
#>   0.105134705  0.491755171  0.129824369   0   0
#>   0.518674769  0.765055442  0.494374052   0   0
#>   0.567659524  0.901814347  0.568374255   0   0
#>   0.620770432  0.873719167  0.654642424   0   0
#>   0.661471473  0.776017980  0.695862821   0   0
#>   0.661527010  0.623961215  0.699478005   0   0
#>   0.581539927  0.630832654  0.540472166   0   0
#>   0.437460252  0.218097252  0.545087378   0   0
#>   0.408046094  0.037333327  0.428739175   0   0
#>   0.349388727  0.176272185  0.320860734   0   0
#>   0.333469795  0.206190633  0.264820841   0   0
#>   0.279052647  0.230082360  0.211674190   0   0
#>   0.289650979  0.399358803  0.282217415   0   0
#>   0.405034334  0.484803787  0.516836392   0   0
#>   0.428910074  0.322555739  0.472772951   0   0
#>   0.180750339  0.450025754  0.166977333   0   0
#>   0.262455243  0.639128393  0.194880848   0   0
#>   0.329255077  0.655179762  0.330450778   0   0
#>   0.340690729  0.562128945  0.399504812   0   0
#>   0.309067012  0.498564348  0.428451547   0   0
#>   0.198986705  0.292058897  0.128425021   0   0
#>   0.677834199  5.825078722  0.663644970   0   0
#>   0.899851770  0.953501368  0.983008103   0   0
#>   0.863928033  0.878859160  0.863550823   0   0
#>   0.799927714  0.928573555  0.772962473   0   0
#>   0.720206574  1.147123430  0.702582857   0   0
#>   0.672088064  0.940107187  0.526804727   0   0
#>   0.211674190  0.206190633  0.264820841   0   0
#>   0.174210830  0.117117916  0.203525119   0   0
#>   0.174132108  0.187292503  0.089070933   0   0
#>   0.128425021  0.294901486  0.107805927   0   0
#>   0.089020985 10.488888974 -6.611065826   0   1
#>   0.252217021 34.090557979  0.746861606   0   0
#>   0.248383974  0.950731326  0.265430193   0   0
#>   0.138064305  0.858288254  0.119518506   0   0
#>   0.399504812  0.571132612  0.463884640   0   0
#>   0.428451547  0.484803787  0.516836392   0   0
#>  -0.245455267  0.464470283  0.147669445   0   0
#>  -0.058076032  0.349096055 -0.008572628   0   0
#>   0.045134653  0.272189548  0.060241004   0   0
#>   0.056288329  0.187292503  0.089070933   0   0
#>  -0.299954994  0.094565108 -0.008747171   0   0
#>   0.104621758  0.812878624  0.024950157   0   0
#>   0.150100328  0.764377834  0.207141915   0   0
#>   0.129824369  0.464470283  0.147669445   0   0
#>   1.323953821  0.662830144  1.311258083   0   0
#>   0.901308405  0.754244059  0.844702327   0   0
#>   0.889960666  0.878859160  0.863550823   0   0
#>   0.068584337  0.349096055 -0.008572628   0   0
#>   0.077886982  0.450025754  0.166977333   0   0
#>   0.024950157  0.858288254  0.119518506   0   0
#>   0.560256171  0.571132612  0.463884640   0   0
#>   0.582322541  0.601634755  0.547688584   0   0
#>   0.649996290  0.477425008  0.662888308   0   0
#>   0.472772951  0.317559509  0.492781490   0   0
#>   0.668624566  0.113428601  0.731046108   0   0
#>   0.573945141  0.026486538  0.693823306   0   0
#>   0.509304925 -0.055106282  0.499429322   0   0
#>   0.552497071  0.218097252  0.545087378   0   0
#>   0.665398585  0.228350195  0.680279596   0   0
#>   0.320860734  0.137166716  0.299616153   0   0
#>   0.282238263  0.148743254  0.274441680   0   0
#>   0.624198560  0.908087010  0.607789277   0   0
#>   0.689793767  1.147123430  0.702582857   0   0
#>   0.722239428  0.928573555  0.772962473   0   0
#>   0.716977307  0.856187226  0.753731773   0   0
#>   0.711407737  0.776017980  0.695862821   0   0
#>   1.054090708 -0.391272096  6.338877857   0   1
#>   0.995837256  0.214891386  0.928690464   0   0
#>   0.984027677  0.365803029  0.983450448   0   0
#>   0.267531481 -2.623832900 -0.033604552   0   0
#>   0.201362926 -0.703325559  0.086046280   0   0
#>   0.191483020  0.070277060  0.120814242   0   0
#>   0.205225751  0.148743254  0.274441680   0   0
#>   0.331755059  0.137166716  0.299616153   0   0
#>   0.399996141  0.037333327  0.428739175   0   0
#>   0.662888308  0.484194886  0.683506125   0   0
#>   0.625502299  0.287046816  0.627437198   0   0
#>   0.492781490  0.245933569  0.568041741   0   0
#>   0.824563565  0.832684997  0.817267187   0   0
#>   0.803002903  0.856187226  0.753731773   0   0
#>   0.840700301  0.657710874  0.934762703   0   0
#>   0.835817970  0.623961215  0.699478005   0   0
#>   0.753311473  0.776017980  0.695862821   0   0
#>   0.775536570  0.856187226  0.753731773   0   0
#>  -0.003003951  0.003777047  0.074827306   0   0
#>   0.074070365  0.187292503  0.089070933   0   0
#>   0.113503792  0.070277060  0.120814242   0   0
#>  -0.033604552 -0.703325559  0.086046280   0   0
#>   0.627437198  0.245933569  0.568041741   0   0
#>   0.683506125  0.623961215  0.699478005   0   0
#>   0.714347530  0.509067773  0.763201843   0   0
#>   0.330450778  0.664358194  0.330753861   0   0
#>   0.430921440  0.668888123  0.498550849   0   0
#>   0.446555974  0.601634755  0.547688584   0   0
#>   0.928690464  0.161540646  0.921237159   0   0
#>   0.746861606  5.825078722  0.663644970   0   0
#>   0.808707647  0.130759754  0.846045575   0   0
#>   0.731046108  0.080938862  0.749085086   0   0
#>   0.545087378  0.245933569  0.568041741   0   0
#>   0.540472166  0.668888123  0.498550849   0   0
#>   0.983450448  0.662830144  1.311258083   0   0
#>   0.958514722  0.644799763  1.224106016   0   0
#>   1.004824917  0.509067773  0.763201843   0   0
#>   1.202437938  0.657710874  0.934762703   0   0
#>   1.224106016  0.662830144  1.311258083   0   0
#>   0.265430193  0.929230977  0.355478211   0   0
#>   0.301627674  0.828055738  0.428433112   0   0
#>   0.280202399  0.664358194  0.330753861   0   0
#>   0.875562441  0.021375320  0.781925186   0   0
#>   0.880270843  0.112760061  0.841890996   0   0
#>   0.911551537  0.132810429  0.867354922   0   0
#>   1.007863125  0.161540646  0.921237159   0   0
#>   0.452999044  0.828055738  0.428433112   0   0
#>   0.274441680  0.137166716  0.299616153   0   0
#>   0.663644970  0.995448418  0.419877184   0   0
#>   0.428433112  0.879531703  0.429420772   0   0
#>   0.166977333  0.457562088  0.154805965   0   0
#>   0.780092866  0.112760061  0.841890996   0   0
#>   0.867354922  0.130759754  0.846045575   0   0
#>   0.526804727  0.925251540  0.486809384   0   0
#>   0.355478211  0.879531703  0.429420772   0   0
#>   0.419877184  0.925251540  0.486809384   0   0
#>   0.486809384  0.879531703  0.429420772   0   0
#>   0.147669445  0.457562088  0.154805965   0   0
#>   0.917719496  0.382163133  0.936952578   0   0
#>   0.936952578  0.354997656  0.920019519   0   0
#>   0.920019519  0.336163195  0.893208907   0   0
#>   0.893208907  0.332647487  0.851785147   0   0
#>   0.851785147  0.379680659  0.796331638   0   0
#>   0.796331638  0.425093939  0.773265701   0   0
#>   0.773265701  0.456111275  0.917719496   0   0
#>   0.856572926  0.198542996  0.803302196   0   0
#>   0.803302196  0.238628952  0.732793100   0   0
#>  -0.288361333  0.640013462  0.087717558   0   0
#>   0.087717558  0.514538909  0.105134705   0   0
#>   0.105134705  0.569892147 -0.288361333   0   0
#>   0.518674769  0.798176800  0.567659524   0   0
#>   0.567659524  0.779430293  0.620770432   0   0
#>   0.620770432  0.753726917  0.661471473   0   0
#>   0.661471473  0.701767364  0.661527010   0   0
#>   0.661527010  0.679929687  0.581539927   0   0
#>   0.581539927  0.771201111  0.518674769   0   0
#>   0.437460252  0.177057689  0.408046094   0   0
#>   0.408046094  0.195265236  0.349388727   0   0
#>   0.349388727  0.208416025  0.333469795   0   0
#>   0.333469795  0.345101470  0.279052647   0   0
#>   0.279052647  0.387104951  0.289650979   0   0
#>   0.289650979  0.344586612  0.405034334   0   0
#>   0.405034334  0.316402019  0.428910074   0   0
#>   0.428910074  0.234056059  0.437460252   0   0
#>   0.180750339  0.628939687  0.262455243   0   0
#>   0.262455243  0.652826317  0.329255077   0   0
#>   0.329255077  0.605608866  0.340690729   0   0
#>   0.340690729  0.441436205  0.309067012   0   0
#>   0.309067012  0.399358803  0.282217415   0   0
#>   0.282217415  0.431230777  0.198986705   0   0
#>   0.198986705  0.444968475  0.180750339   0   0
#>   0.677834199  0.952317674  0.899851770   0   0
#>   0.899851770  0.933601674  0.863928033   0   0
#>   0.863928033  0.985156943  0.799927714   0   0
#>   0.799927714  1.260376947  0.720206574   0   0
#>   0.720206574  1.856251061  0.672088064   0   0
#>   0.672088064  2.531309820  0.677834199   0   0
#>   0.211674190  0.203645528  0.174210830   0   0
#>   0.174210830  0.203668271  0.174132108   0   0
#>   0.174132108  0.292058897  0.128425021   0   0
#>   0.089020985  1.246526191  0.252217021   0   0
#>   0.252217021  0.965723070  0.248383974   0   0
#>   0.248383974  0.887283827  0.138064305   0   0
#>   0.138064305  0.958303647  0.089020985   0   0
#>   0.399504812  0.498564348  0.428451547   0   0
#>  -2.191867715  0.375933323 -0.245455267   0   0
#>  -0.245455267  0.338300917 -0.058076032   0   0
#>  -0.058076032  0.260630512  0.045134653   0   0
#>   0.045134653  0.215346451  0.056288329   0   0
#>   0.056288329  0.111651923 -0.299954994   0   0
#>  -0.299954994 -0.112161239 -4.892386873   0   1
#>   0.104621758  0.670246335  0.150100328   0   0
#>   0.150100328  0.639128393  0.194880848   0   0
#>   0.194880848  0.491755171  0.129824369   0   0
#>   1.323953821  0.845868677  0.901308405   0   0
#>   0.901308405  0.860750211  0.889960666   0   0
#>   0.889960666  0.953501368  0.983008103   0   0
#>   0.983008103  3.039810863  9.705536906   0   1
#>   0.060241004  0.337162434  0.068584337   0   0
#>   0.068584337  0.338401463  0.077886982   0   0
#>   0.077886982  0.294901486  0.107805927   0   0
#>   0.107805927  0.272189548  0.060241004   0   0
#>   0.560256171  0.504994385  0.582322541   0   0
#>   0.582322541  0.480105207  0.649996290   0   0
#>   0.649996290  0.322555739  0.472772951   0   0
#>   0.516836392  0.502392418  0.560256171   0   0
#>   0.668624566  0.011572850  0.573945141   0   0
#>   0.573945141 -0.042633469  0.509304925   0   0
#>   0.509304925  0.192546287  0.552497071   0   0
#>   0.552497071  0.162128782  0.665398585   0   0
#>   0.665398585  0.126722862  0.668624566   0   0
#>   0.320860734  0.166924026  0.282238263   0   0
#>   0.282238263  0.206190633  0.264820841   0   0
#>   0.624198560  0.941472124  0.689793767   0   0
#>   0.689793767  0.904094515  0.722239428   0   0
#>   0.722239428  0.863147165  0.716977307   0   0
#>   0.716977307  0.857397315  0.711407737   0   0
#>   0.711407737  0.873719167  0.654642424   0   0
#>   0.654642424  0.905454625  0.624198560   0   0
#>   1.054090708  0.255985706  0.995837256   0   0
#>   0.984027677  0.263041685  1.054090708   0   0
#>   0.267531481  0.056580762  0.201362926   0   0
#>   0.201362926  0.080894456  0.191483020   0   0
#>   0.191483020  0.117117916  0.203525119   0   0
#>   0.203525119  0.117264856  0.205225751   0   0
#>   0.205225751  0.087139937  0.331755059   0   0
#>   0.331755059  0.045136458  0.399996141   0   0
#>   0.399996141  0.036623621  0.267531481   0   0
#>   0.662888308  0.323501103  0.625502299   0   0
#>   0.625502299  0.317559509  0.492781490   0   0
#>   0.863550823  0.861883519  0.824563565   0   0
#>   0.824563565  0.871319621  0.803002903   0   0
#>   0.803002903  0.928573555  0.772962473   0   0
#>   0.844702327  0.713482491  0.840700301   0   0
#>   0.840700301  0.713391029  0.835817970   0   0
#>   0.835817970  0.769803263  0.753311473   0   0
#>   0.753311473  0.812063023  0.775536570   0   0
#>   0.775536570  0.832684997  0.817267187   0   0
#>   0.817267187  0.754244059  0.844702327   0   0
#>  -0.003003951  0.094565108 -0.008747171   0   0
#>  -0.008747171  0.152621288  0.074070365   0   0
#>   0.074070365  0.077734028  0.113503792   0   0
#>   0.113503792  0.090165380 -0.003003951   0   0
#>   0.428739175 -0.055106282  0.499429322   0   0
#>   0.499429322 -2.302817967  0.512317175   0   1
#>   0.680279596  0.287046816  0.627437198   0   0
#>   0.683506125  0.475815377  0.714347530   0   0
#>   0.330450778  0.638939004  0.430921440   0   0
#>   0.430921440  0.615630268  0.446555974   0   0
#>   0.446555974  0.571132612  0.463884640   0   0
#>   0.808707647  0.113428601  0.731046108   0   0
#>   0.074827306  0.070277060  0.120814242   0   0
#>   0.086046280  0.003777047  0.074827306   0   0
#>   0.547688584  0.630832654  0.540472166   0   0
#>   0.958514722  0.365803029  0.983450448   0   0
#>   1.004824917  0.641659549  1.202437938   0   0
#>   1.202437938  0.644799763  1.224106016   0   0
#>   0.265430193  0.810892192  0.301627674   0   0
#>   0.301627674  0.752774466  0.280202399   0   0
#>   0.280202399  0.764377834  0.207141915   0   0
#>   0.207141915  0.858288254  0.119518506   0   0
#>   0.875562441  0.006788834  0.880270843   0   0
#>   0.880270843  0.076611942  0.911551537   0   0
#>   0.911551537  0.094999086  1.007863125   0   0
#>   1.007863125 -2.639767728  8.052561671   0   1
#>   0.330753861  0.779663414  0.452999044   0   0
#>   0.452999044  0.765055442  0.494374052   0   0
#>   0.494374052  0.668888123  0.498550849   0   0
#>   0.763201843  0.657710874  0.934762703   0   0
#>   0.749085086  0.041378890  0.780092866   0   0
#>   0.780092866  0.021375320  0.781925186   0   0
#>   0.781925186  0.026486538  0.693823306   0   0
#>   0.693823306  0.080938862  0.749085086   0   0
#>   0.867354922  0.161540646  0.921237159   0   0
#>   0.568374255  0.940107187  0.526804727   0   0
#>   0.702582857  0.908087010  0.607789277   0   0
#>   0.607789277  0.901814347  0.568374255   0   0
#>   0.419877184  0.929230977  0.355478211   0   0
#>   0.846045575  0.112760061  0.841890996   0   0
#>   0.154805965  0.349096055 -0.008572628   0   0
#>  -2.191867715  0.695625908 -4.463499363   0   1
#>   1.323953821  0.883490955  3.216300535   0   1
#>   0.024950157  1.051264200 -3.575249074   0   1
#>  -0.033604552 -6.365175757 -0.285459152   0   1
#>   0.746861606 35.337528724  0.755668981   0   1
#>   0.875562441 -4.705657909  2.013433896   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.400948620 0.90705319
#>   [2,] 0.266972559 0.79342073
#>   [3,] 0.585139214 0.04022252
#>   [4,] 0.716805733 0.58170752
#>   [5,] 0.254173502 0.34986118
#>   [6,] 0.546868194 0.26863166
#>   [7,] 0.988784907 0.87459371
#>   [8,] 0.298652542 0.17358113
#>   [9,] 0.999606101 0.11286642
#>  [10,] 0.524303772 0.38577352
#>  [11,] 0.235145072 0.00000000
#>  [12,] 0.600150232 0.14836312
#>  [13,] 0.885894451 0.94001195
#>  [14,] 0.309009156 0.07307424
#>  [15,] 0.674084279 0.02363838
#>  [16,] 0.384144478 0.50328800
#>  [17,] 0.158489001 0.59077979
#>  [18,] 0.183683104 0.29162835
#>  [19,] 0.897793164 0.67668015
#>  [20,] 0.327216356 0.95885013
#>  [21,] 0.074115065 0.27190189
#>  [22,] 0.345646778 0.53751200
#>  [23,] 0.911430666 0.81228123
#>  [24,] 0.800152301 0.79818414
#>  [25,] 0.886478556 0.76472510
#>  [26,] 0.095854042 0.07029665
#>  [27,] 0.449181841 0.42172103
#>  [28,] 0.000000000 0.27666490
#>  [29,] 0.302817263 0.71384765
#>  [30,] 0.586347667 0.43164117
#>  [31,] 0.274637595 0.91777343
#>  [32,] 1.000000000 0.11217075
#>  [33,] 0.172034642 0.73944645
#>  [34,] 0.823973755 0.75288900
#>  [35,] 0.019942336 0.13858695
#>  [36,] 0.841616454 0.66052754
#>  [37,] 0.174314380 0.50461114
#>  [38,] 0.648254753 0.60042297
#>  [39,] 0.202762036 0.60270766
#>  [40,] 0.363591254 0.96698511
#>  [41,] 0.819137974 0.85246686
#>  [42,] 0.069138552 0.06744611
#>  [43,] 0.413135669 0.95391046
#>  [44,] 0.826594588 0.23588068
#>  [45,] 0.054624106 0.92903723
#>  [46,] 0.688625178 0.44817400
#>  [47,] 0.547655673 0.43705233
#>  [48,] 0.171211789 0.26351246
#>  [49,] 0.002968025 0.25044319
#>  [50,] 0.168431644 0.01941880
#>  [51,] 0.994336143 0.91407418
#>  [52,] 0.780161371 0.36183439
#>  [53,] 0.530385443 0.88367918
#>  [54,] 0.017911473 0.01058693
#>  [55,] 0.332381530 0.10705598
#>  [56,] 0.156313864 0.29825278
#>  [57,] 0.047055641 0.72412913
#>  [58,] 0.103495036 0.91970713
#>  [59,] 0.271741561 0.51905751
#>  [60,] 0.692918611 0.54702676
#>  [61,] 0.967531225 0.61139937
#>  [62,] 0.995894455 0.38477589
#>  [63,] 0.002550851 0.72154719
#>  [64,] 0.910721047 0.91526440
#>  [65,] 0.056485565 0.71483733
#>  [66,] 0.992357830 0.45484500
#>  [67,] 0.177984329 0.97692593
#>  [68,] 0.464586561 0.02326433
#>  [69,] 0.364349889 1.00000000
#>  [70,] 0.125738959 0.28419285
#>  [71,] 0.097078538 0.78794874
#>  [72,] 0.852700151 0.62967405
#>  [73,] 0.604146575 0.47734634
#>  [74,] 0.618320495 0.47930784
#>  [75,] 0.625281126 0.80146003
#>  [76,] 0.696318575 0.21519042
#>  [77,] 0.950634189 0.04195001
#>  [78,] 0.411034805 0.03532491
#>  [79,] 0.790228455 0.89926084
#>  [80,] 0.230387120 0.29011435
#>  [81,] 0.057128382 0.83409397
#>  [82,] 0.075023163 0.88350354
#>  [83,] 0.948315664 0.73488240
#>  [84,] 0.502032840 0.88534824
#>  [85,] 0.293581543 0.53984284
#>  [86,] 0.382415275 0.98363092
#>  [87,] 0.853550284 0.50640290
#>  [88,] 0.578492400 0.19742487
#>  [89,] 0.740012236 0.14308601
#>  [90,] 0.856522406 0.35149867
#>  [91,] 0.365509437 0.06554892
#>  [92,] 0.953849294 0.64590044
#>  [93,] 0.837288444 0.77983262
#>  [94,] 0.146443590 0.31642126
#>  [95,] 0.311056991 0.05712693
#>  [96,] 0.716976327 0.74131417
#>  [97,] 0.123612094 0.12301143
#>  [98,] 0.192616751 0.87218709
#>  [99,] 0.960373145 0.42130023
#> [100,] 0.259022478 0.09694256
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.4009486,0.9070532) [7]: 2 20 29 31 40 43 84 
#> 2: (0.2669726,0.7934207) [5]: 1 29 31 33 98 
#> 3: (0.5851392,0.04022252) [3]: 12 15 68 
#> 4: (0.7168057,0.5817075) [6]: 36 38 60 72 87 96 
#> 5: (0.2541735,0.3498612) [8]: 8 16 18 27 37 59 80 94 
#> 6: (0.5468682,0.2686317) [7]: 8 10 27 30 55 76 88 
#> 7: (0.9887849,0.8745937) [6]: 23 51 61 64 66 83 
#> 8: (0.2986525,0.1735811) [8]: 5 6 27 48 55 80 97 100 
#> 9: (0.9996061,0.1128664) [4]: 32 44 62 77 
#> 10: (0.5243038,0.3857735) [4]: 6 27 30 47 
#> 11: (0.2351451,0) [7]: 15 50 54 68 78 95 100 
#> 12: (0.6001502,0.1483631) [6]: 3 15 68 76 88 89 
#> 13: (0.8858945,0.9400119) [5]: 41 51 64 69 79 
#> 14: (0.3090092,0.07307424) [4]: 55 91 95 100 
#> 15: (0.6740843,0.02363838) [6]: 3 11 12 68 77 89 
#> 16: (0.3841445,0.503288) [7]: 5 22 27 38 47 59 73 
#> 17: (0.158489,0.5907798) [5]: 33 37 39 63 65 
#> 18: (0.1836831,0.2916283) [5]: 5 48 56 80 94 
#> 19: (0.8977932,0.6766802) [6]: 25 34 36 72 83 92 
#> 20: (0.3272164,0.9588501) [5]: 1 31 40 67 69 
#> 21: (0.07411506,0.2719019) [7]: 28 35 48 49 70 94 97 
#> 22: (0.3456468,0.537512) [5]: 16 29 38 59 85 
#> 23: (0.9114307,0.8122812) [6]: 7 25 41 64 83 93 
#> 24: (0.8001523,0.7981841) [6]: 34 41 75 79 93 96 
#> 25: (0.8864786,0.7647251) [5]: 19 23 34 83 93 
#> 26: (0.09585404,0.07029665) [4]: 42 50 54 97 
#> 27: (0.4491818,0.421721) [6]: 5 6 8 10 16 47 
#> 28: (0,0.2766649) [6]: 21 37 49 54 63 94 
#> 29: (0.3028173,0.7138477) [9]: 1 2 22 33 38 39 75 84 85 
#> 30: (0.5863477,0.4316412) [7]: 6 10 46 47 73 74 76 
#> 31: (0.2746376,0.9177734) [5]: 1 2 20 67 98 
#> 32: (1,0.1121708) [4]: 9 51 62 77 
#> 33: (0.1720346,0.7394465) [7]: 2 17 29 39 65 71 98 
#> 34: (0.8239738,0.752889) [6]: 19 24 25 36 93 96 
#> 35: (0.01994234,0.1385869) [5]: 21 42 49 54 97 
#> 36: (0.8416165,0.6605275) [5]: 4 19 34 72 96 
#> 37: (0.1743144,0.5046111) [7]: 5 17 28 39 59 63 94 
#> 38: (0.6482548,0.600423) [9]: 4 16 22 29 60 73 74 75 96 
#> 39: (0.202762,0.6027077) [6]: 17 29 33 37 59 85 
#> 40: (0.3635913,0.9669851) [5]: 1 20 43 69 86 
#> 41: (0.819138,0.8524669) [6]: 13 23 24 64 79 93 
#> 42: (0.06913855,0.06744611) [4]: 26 35 54 97 
#> 43: (0.4131357,0.9539105) [6]: 1 40 53 79 84 86 
#> 44: (0.8265946,0.2358807) [7]: 9 52 62 76 77 89 90 
#> 45: (0.05462411,0.9290372) [5]: 58 63 67 81 82 
#> 46: (0.6886252,0.448174) [6]: 30 52 60 74 76 87 
#> 47: (0.5476557,0.4370523) [5]: 10 16 27 30 73 
#> 48: (0.1712118,0.2635125) [7]: 8 18 21 56 70 80 97 
#> 49: (0.002968025,0.2504432) [4]: 21 28 35 54 
#> 50: (0.1684316,0.0194188) [5]: 11 26 54 97 100 
#> 51: (0.9943361,0.9140742) [6]: 7 13 32 62 64 66 
#> 52: (0.7801614,0.3618344) [5]: 44 46 76 87 90 
#> 53: (0.5303854,0.8836792) [4]: 43 75 79 84 
#> 54: (0.01791147,0.01058693) [7]: 11 26 28 35 42 49 50 
#> 55: (0.3323815,0.107056) [6]: 6 8 14 88 91 100 
#> 56: (0.1563139,0.2982528) [4]: 18 48 70 94 
#> 57: (0.04705564,0.7241291) [4]: 63 65 71 81 
#> 58: (0.103495,0.9197071) [4]: 45 67 82 98 
#> 59: (0.2717416,0.5190575) [6]: 5 16 22 37 39 85 
#> 60: (0.6929186,0.5470268) [5]: 4 38 46 74 87 
#> 61: (0.9675312,0.6113994) [6]: 7 66 72 83 87 92 
#> 62: (0.9958945,0.3847759) [7]: 9 32 44 51 66 90 99 
#> 63: (0.002550851,0.7215472) [7]: 17 28 37 45 57 65 81 
#> 64: (0.910721,0.9152644) [5]: 7 13 23 41 51 
#> 65: (0.05648557,0.7148373) [5]: 17 33 57 63 71 
#> 66: (0.9923578,0.454845) [6]: 7 51 61 62 87 99 
#> 67: (0.1779843,0.9769259) [6]: 20 31 45 58 69 98 
#> 68: (0.4645866,0.02326433) [6]: 3 11 12 15 78 88 
#> 69: (0.3643499,1) [6]: 13 20 40 67 79 86 
#> 70: (0.125739,0.2841928) [4]: 21 48 56 94 
#> 71: (0.09707854,0.7879487) [6]: 33 57 65 81 82 98 
#> 72: (0.8527002,0.629674) [6]: 4 19 36 61 87 92 
#> 73: (0.6041466,0.4773463) [5]: 16 30 38 47 74 
#> 74: (0.6183205,0.4793078) [5]: 30 38 46 60 73 
#> 75: (0.6252811,0.80146) [7]: 24 29 38 53 79 84 96 
#> 76: (0.6963186,0.2151904) [8]: 6 12 30 44 46 52 88 89 
#> 77: (0.9506342,0.04195001) [5]: 9 15 32 44 89 
#> 78: (0.4110348,0.03532491) [5]: 11 68 88 91 95 
#> 79: (0.7902285,0.8992608) [8]: 13 24 41 43 53 69 75 86 
#> 80: (0.2303871,0.2901143) [4]: 5 8 18 48 
#> 81: (0.05712838,0.834094) [5]: 45 57 63 71 82 
#> 82: (0.07502316,0.8835035) [5]: 45 58 71 81 98 
#> 83: (0.9483157,0.7348824) [6]: 7 19 23 25 61 92 
#> 84: (0.5020328,0.8853482) [5]: 1 29 43 53 75 
#> 85: (0.2935815,0.5398428) [4]: 22 29 39 59 
#> 86: (0.3824153,0.9836309) [4]: 40 43 69 79 
#> 87: (0.8535503,0.5064029) [9]: 4 46 52 60 61 66 72 90 99 
#> 88: (0.5784924,0.1974249) [7]: 6 12 55 68 76 78 91 
#> 89: (0.7400122,0.143086) [5]: 12 15 44 76 77 
#> 90: (0.8565224,0.3514987) [5]: 44 52 62 87 99 
#> 91: (0.3655094,0.06554892) [5]: 14 55 78 88 95 
#> 92: (0.9538493,0.6459004) [4]: 19 61 72 83 
#> 93: (0.8372884,0.7798326) [5]: 23 24 25 34 41 
#> 94: (0.1464436,0.3164213) [7]: 5 18 21 28 37 56 70 
#> 95: (0.311057,0.05712693) [5]: 11 14 78 91 100 
#> 96: (0.7169763,0.7413142) [6]: 4 24 34 36 38 75 
#> 97: (0.1236121,0.1230114) [8]: 8 21 26 35 42 48 50 100 
#> 98: (0.1926168,0.8721871) [7]: 2 31 33 58 67 71 82 
#> 99: (0.9603731,0.4213002) [4]: 62 66 87 90 
#> 100: (0.2590225,0.09694256) [7]: 8 11 14 50 55 95 97 
#> number of nodes: 100 
#> number of arcs: 285 
#> number of boundary nodes: 12 
#> boundary nodes:  11 13 15 28 32 45 51 54 63 67 69 77 
#> number of triangles: 186 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                               
#> 0.1122652451 0.0512405656 0.0372734647 0.0667219659 0.1245887174 0.0872738029 
#>                                                                               
#> 0.2628674498 0.0938378060 0.1343268355 0.1332643711 0.2094980519 0.1449913973 
#>                                                                               
#> 0.1844646760 0.1656889640 0.1232740814 0.0327839360 0.1485316124 0.1567577216 
#>                                                                               
#> 0.0696133294 0.0984924017 0.1902425047 0.0447377897 0.0467285495 0.1350561279 
#>                                                                               
#> 0.2902283206 0.1043217598 0.1135037145 0.2621776051 0.1191579713 0.2427849430 
#>                                                                               
#> 0.0771575404 0.0832798157 0.0745871318 0.4592334414 0.0836235666 0.1029856197 
#>                                                                               
#> 0.0857384543 0.1249691746 0.1585106688 0.0648797343 0.1483451112 0.1378969122 
#>                                                                               
#> 0.0740528988 0.0858366822 0.2726360518 0.2254600372 0.2302062720 0.0390685433 
#>                                                                               
#> 0.0996601450 0.0548930598 0.1023273745 0.0655200323 0.1192335189 0.1507790569 
#>                                                                               
#> 0.1364339853 0.0843100315 0.2081019780 0.4376310446 0.0550039713 0.1110449939 
#>                                                                               
#> 0.0550998973 0.0531064482 0.2336452405 0.0693889703 0.1307416833 0.3090782691 
#>                                                                               
#> 0.0761744672 0.1181405865 0.0543504883 0.2767019341 0.1021381293 0.1401487424 
#>                                                                               
#> 0.0206764561 0.0377999731 0.1024423982 0.0891538517 0.0686615990 0.0636155917 
#>                                                                               
#> 0.0940313990 0.1877885374 0.1133175692 0.0330236048 0.0263891478 0.1131368473 
#>                                                                               
#> 0.1048332739 0.0499545433 0.0383060360 0.1517441712 0.3635824773 0.1742497454 
#>                                                                               
#> 0.0301498721 0.0748677041 0.0514577894 0.1917618846 0.1096609986 0.1076216823 
#>                                                                               
#> 0.0300539395 0.0765322202 0.1128726056 0.0778130520 0.2403213127 0.1104543735 
#>                                                                               
#> 0.2023454607 0.1490884688 0.2331105689 0.0768899511 0.0143090008 0.1057560060 
#>                                                                               
#> 0.5293005848 0.1273720913 0.0836245625 0.1084228892 0.1007508353 0.0243783707 
#>                                                                               
#> 0.0427439866 0.0284016872 0.2603097751 0.4164491566 0.1432897114 0.0770573480 
#>                                                                               
#> 0.1689203072 0.1250819111 0.0525502517 0.0460580887 0.0939289395 0.1621295313 
#>                                                                               
#> 0.0336527355 0.0701583066 0.1549327387 0.2505055095 0.0610360317 0.1181368449 
#>                                                                               
#> 0.1480734668 0.1251288708 0.0463494488 0.1365781356 0.2330631771 0.0484161902 
#>                                                                               
#> 0.0706215793 0.0901075412 0.1267651321 0.1756756167 0.2166982778 0.1033882229 
#>                                                                               
#> 0.1083186305 0.1092081970 0.0904779438 0.1091774751 0.1217395673 0.1561084846 
#>                                                                               
#> 0.1441113474 0.1476154178 0.1596067346 0.0710598682 0.0421111611 0.1741408328 
#>                                                                               
#> 0.1128005401 0.0914328073 0.0643076640 0.1818049473 0.2078270697 0.2010777093 
#>                                                                               
#> 0.1701059553 0.0779133932 0.1587179334 0.1677221818 0.1192953048 0.1816011543 
#>                                                                               
#> 0.2657924229 0.2685353379 0.0398688299 0.0880231302 0.0993303792 0.1454544999 
#>                                                                               
#> 0.2640510915 0.4197639181 0.1559768895 0.1821989390 0.0862786960 0.0007994458 
#>                                                                               
#> 0.2719347991 0.2122863422 0.0861822823 0.0563456111 0.2306179225 0.1794019155 
#>                                                                               
#> 0.0950058183 0.0998398197 0.0694821646 0.2174914243 0.1399615236 0.1171078022 
#>                                                                               
#> 0.0536294376 0.1039838232 0.1100934576 0.0350542560 0.1115005315 0.0160782619 
#>                                                                               
#> 0.0569992308 0.0412435000 0.0553928174 0.2215262783 0.2814061833 0.0515107303 
#>                                                                               
#> 0.1764173303 0.1606081410 0.2035112282 0.0876098098 0.0458516726 0.1492824926 
#>                                                                               
#> 0.0281595179 0.0307577094 0.0639506090 0.0770715513 0.0887689912 0.1060994526 
#>                                                                               
#> 0.0584527980 0.0651379589 0.1503227614 0.0554275330 0.0743126995 0.1439012154 
#>                                                                               
#> 0.1569022837 0.0974584854 0.0530668824 0.0849316755 0.0742679540 0.1814624725 
#>                                                                               
#> 0.0521173819 0.1006619391 0.0809319560 0.0537046808 0.1015626981 0.1749018570 
#>                                                                               
#> 0.1007593024 0.0511772552 0.0414230807 0.0575071236 0.0981850054 0.0886344405 
#>                                                                               
#> 0.0595764826 0.0268671361 0.2869581705 0.4448896044 0.1495431250 0.3341539625 
#>                                                                               
#> 0.1036051341 0.0573966316 0.0490485891 0.0892798353 0.0864944264 0.1280161301 
#>                                                                               
#> 0.1247595077 0.0251282061 0.1366745841 0.3810322166 0.1194286150 0.1342400172 
#>                                                                               
#> 0.1319087811 0.1269147503 0.0949762846 0.0498942735 0.0497535778 0.1323294065 
#>                                                                               
#> 0.1258308460 0.1749025306 0.0989459549 0.1255594666 0.0810878099 0.1104252065 
#>                                                                               
#> 0.0445796223 0.0132386169 0.1009991814 0.1549706890 0.0371149421 0.1162761339 
#>                                                                               
#> 0.0509489077 0.0980671185 0.0546447612 0.4395752504 0.5249831392 0.2771554947 
#>                                        
#> 0.2666801588 0.8019234306 0.2139246169 
#> 
#> $alpha
#> [1] 0.2892633
#> 
#> attr(,"class")
#> [1] "scree"