Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1
#>    60   91 0.87089760 0.84224327 0.91338248 0.83220498  0.915768680
#>    60    1 0.87089760 0.84224327 1.00000000 1.00000000  0.884681054
#>    19   55 0.90621106 0.79076398 0.93790499 0.72394215  0.947509154
#>    84   19 0.81369490 0.75866931 0.90621106 0.79076398  0.853741201
#>    65   84 0.77936952 0.90265472 0.81369490 0.75866931  0.797743395
#>    71    1 0.95682906 0.72388305 1.00000000 1.00000000  1.097758611
#>    70   55 0.81812899 0.71239631 0.93790499 0.72394215  0.876330551
#>    91   19 0.91338248 0.83220498 0.90621106 0.79076398  0.985721924
#>    70   90 0.81812899 0.71239631 0.82042990 0.67676730  0.879913253
#>    84   70 0.81369490 0.75866931 0.81812899 0.71239631  0.871691939
#>    95   71 0.86109722 0.57878841 0.95682906 0.72388305  0.947061870
#>    84   27 0.81369490 0.75866931 0.67907927 0.71784563  0.742233553
#>    10   96 0.62297677 0.74842619 0.63303302 0.72117683  0.657962130
#>    22   10 0.59991347 0.78680052 0.62297677 0.74842619  0.698262485
#>    65   22 0.77936952 0.90265472 0.59991347 0.78680052  0.702556984
#>    27   96 0.67907927 0.71784563 0.63303302 0.72117683  0.653341094
#>    94   96 0.62714701 0.64644988 0.63303302 0.72117683  0.615053051
#>    90   95 0.82042990 0.67676730 0.86109722 0.57878841  0.898661989
#>    90   25 0.82042990 0.67676730 0.71172367 0.68698470  0.757753664
#>    13   95 0.84624093 0.50671631 0.86109722 0.57878841  0.738342362
#>    65    1 0.77936952 0.90265472 1.00000000 1.00000000  0.769771796
#>    13   25 0.84624093 0.50671631 0.71172367 0.68698470  0.732166793
#>    25   27 0.71172367 0.68698470 0.67907927 0.71784563  0.666817550
#>    22    5 0.59991347 0.78680052 0.57584425 0.69473807  0.573175989
#>    22   51 0.59991347 0.78680052 0.55249653 0.98736727  0.571851668
#>    13   11 0.84624093 0.50671631 0.99584295 0.53866786  0.919662892
#>    45    5 0.50529775 0.76746954 0.57584425 0.69473807  0.558464305
#>    45   22 0.50529775 0.76746954 0.59991347 0.78680052  0.539714875
#>    49   51 0.46897473 0.87862548 0.55249653 0.98736727  0.527887355
#>    67   42 0.45600321 0.91906964 0.44289973 0.94738632  0.438979939
#>    49   67 0.46897473 0.87862548 0.45600321 0.91906964  0.434592751
#>    67   51 0.45600321 0.91906964 0.55249653 0.98736727  0.501388170
#>    94    5 0.62714701 0.64644988 0.57584425 0.69473807  0.558427719
#>     5   82 0.57584425 0.69473807 0.50574640 0.68177240  0.552032418
#>    45   49 0.50529775 0.76746954 0.46897473 0.87862548  0.418600329
#>    44   49 0.41985445 0.70782967 0.46897473 0.87862548  0.400096752
#>     9   28 0.48053435 0.65840022 0.52104626 0.56325761  0.523481687
#>    82   45 0.50574640 0.68177240 0.50529775 0.76746954  0.481910964
#>    82    4 0.50574640 0.68177240 0.49161032 0.66809140  0.545408205
#>    28   69 0.52104626 0.56325761 0.65064580 0.41894810  0.631288441
#>    47    9 0.44331640 0.69443060 0.48053435 0.65840022  0.468556233
#>    31   13 0.94379629 0.43323590 0.84624093 0.50671631  0.896926717
#>    48   13 0.90574145 0.41199935 0.84624093 0.50671631  0.808726496
#>    25   94 0.71172367 0.68698470 0.62714701 0.64644988  0.719706381
#>    47   44 0.44331640 0.69443060 0.41985445 0.70782967  0.377389833
#>    47   45 0.44331640 0.69443060 0.50529775 0.76746954  0.457002147
#>    11   31 0.99584295 0.53866786 0.94379629 0.43323590  1.047186034
#>    31   48 0.94379629 0.43323590 0.90574145 0.41199935  0.935769481
#>    80   43 0.71354667 0.43656396 0.71114800 0.20604155  0.796370472
#>    33   92 0.34345086 0.95865233 0.31205899 0.75830901  0.342705921
#>    59   42 0.42029734 0.92314035 0.44289973 0.94738632  0.395026055
#>    42   51 0.44289973 0.94738632 0.55249653 0.98736727  0.419183546
#>    40   30 0.84457203 0.16187190 0.98485155 0.30473233  0.855095114
#>    69   80 0.65064580 0.41894810 0.71354667 0.43656396  0.652317911
#>    54    9 0.38460384 0.48796418 0.48053435 0.65840022  0.401019373
#>    49   59 0.46897473 0.87862548 0.42029734 0.92314035  0.380825955
#>    11   98 0.99584295 0.53866786 0.96257273 0.38659634  1.214421540
#>    98   48 0.96257273 0.38659634 0.90574145 0.41199935  0.901385184
#>    35   34 0.26775950 0.99579311 0.20157168 0.94781712  0.275383021
#>    33   51 0.34345086 0.95865233 0.55249653 0.98736727  0.416383833
#>    51    1 0.55249653 0.98736727 1.00000000 1.00000000  0.597531534
#>    89   58 0.57696116 0.34830324 0.60409908 0.30807945  0.636934614
#>    89   69 0.57696116 0.34830324 0.65064580 0.41894810  0.547290699
#>    54   81 0.38460384 0.48796418 0.47200612 0.37442518  0.482082362
#>    69   58 0.65064580 0.41894810 0.60409908 0.30807945  0.709776731
#>    89   81 0.57696116 0.34830324 0.47200612 0.37442518  0.530209224
#>     6   73 0.16502639 0.95756678 0.16138271 0.93557897  0.180884928
#>    34   92 0.20157168 0.94781712 0.31205899 0.75830901  0.279166136
#>    54   72 0.38460384 0.48796418 0.35138810 0.49923270  0.353781187
#>     2   72 0.33115035 0.50051474 0.35138810 0.49923270  0.348605433
#>    21   89 0.54552234 0.31722889 0.57696116 0.34830324  0.545557772
#>    72  100 0.35138810 0.49923270 0.33241176 0.49514781  0.351228127
#>    92    2 0.31205899 0.75830901 0.33115035 0.50051474  0.317264639
#>    85   66 0.52070102 0.28728793 0.52417624 0.32563156  0.462429219
#>    52   92 0.14511855 0.80699219 0.31205899 0.75830901  0.247546577
#>    73   52 0.16138271 0.93557897 0.14511855 0.80699219  0.204908528
#>    78    2 0.16944491 0.64335481 0.33115035 0.50051474  0.299631766
#>    58   43 0.60409908 0.30807945 0.71114800 0.20604155  0.718794634
#>    58   85 0.60409908 0.30807945 0.52070102 0.28728793  0.585178990
#>    58   21 0.60409908 0.30807945 0.54552234 0.31722889  0.568592594
#>    43   40 0.71114800 0.20604155 0.84457203 0.16187190  0.816684700
#>    52   78 0.14511855 0.80699219 0.16944491 0.64335481  0.214299132
#>    77   85 0.39075448 0.29857242 0.52070102 0.28728793  0.457117174
#>   100    2 0.33241176 0.49514781 0.33115035 0.50051474  0.287650048
#>    23   78 0.09770875 0.75226346 0.16944491 0.64335481  0.180320693
#>    77   37 0.39075448 0.29857242 0.39912401 0.24399954  0.449257291
#>    20   77 0.30651168 0.44614346 0.39075448 0.29857242  0.380774642
#>    99   77 0.38012586 0.28949722 0.39075448 0.29857242  0.342877681
#>    63   46 0.16290926 0.47673770 0.20083599 0.45829540  0.225730329
#>    83   79 0.11962555 0.91535497 0.03768689 0.89191356  0.093424850
#>    85   15 0.52070102 0.28728793 0.45798725 0.21874199  0.551319492
#>    53   29 0.28048834 0.36516186 0.28492304 0.28644871  0.331225478
#>    20    2 0.30651168 0.44614346 0.33115035 0.50051474  0.259208219
#>    16   53 0.21383728 0.43483484 0.28048834 0.36516186  0.263238789
#>    46   16 0.20083599 0.45829540 0.21383728 0.43483484  0.256133259
#>    99   37 0.38012586 0.28949722 0.39912401 0.24399954  0.403714975
#>    79   75 0.03768689 0.89191356 0.00000000 0.82756695  0.073999283
#>    23   52 0.09770875 0.75226346 0.14511855 0.80699219  0.074331293
#>    83    6 0.11962555 0.91535497 0.16502639 0.95756678  0.036792612
#>    36    7 0.09859034 0.58362591 0.11980498 0.51739171  0.170545562
#>    29   37 0.28492304 0.28644871 0.39912401 0.24399954  0.333948533
#>    16   26 0.21383728 0.43483484 0.13346935 0.42717367  0.182109000
#>    36   78 0.09859034 0.58362591 0.16944491 0.64335481  0.095693314
#>     7   63 0.11980498 0.51739171 0.16290926 0.47673770  0.199524738
#>    43   93 0.71114800 0.20604155 0.57504499 0.10330154  0.689293967
#>    74   41 0.40337241 0.06109896 0.42411968 0.01316131  0.491966123
#>    18   36 0.05636180 0.58763117 0.09859034 0.58362591  0.083106800
#>     8   23 0.05962720 0.72899873 0.09770875 0.75226346  0.063285743
#>    76   37 0.22994046 0.23636208 0.39912401 0.24399954  0.316501328
#>    14   75 0.02630406 0.76670757 0.00000000 0.82756695 -0.164979608
#>    52   75 0.14511855 0.80699219 0.00000000 0.82756695  0.073395117
#>    26   63 0.13346935 0.42717367 0.16290926 0.47673770  0.120637028
#>    78    8 0.16944491 0.64335481 0.05962720 0.72899873  0.095935913
#>    24   88 0.00108610 0.58600821 0.01670961 0.56077328  0.028757235
#>    14   17 0.02630406 0.76670757 0.02046862 0.65406897  0.001833491
#>    41   38 0.42411968 0.01316131 0.60552041 0.01516878  0.514610850
#>    57   29 0.24659482 0.26874459 0.28492304 0.28644871  0.244485598
#>    26   97 0.13346935 0.42717367 0.10549624 0.42796597  0.116543055
#>    39   97 0.05966839 0.48888503 0.10549624 0.42796597  0.103479896
#>    18   39 0.05636180 0.58763117 0.05966839 0.48888503  0.072853270
#>    57   76 0.24659482 0.26874459 0.22994046 0.23636208  0.289496596
#>    17   24 0.02046862 0.65406897 0.00108610 0.58600821  0.027891705
#>    24   62 0.00108610 0.58600821 0.01738048 0.49252307 -0.067949690
#>    75   24 0.00000000 0.82756695 0.00108610 0.58600821 -0.289261928
#>    37   74 0.39912401 0.24399954 0.40337241 0.06109896  0.394668635
#>    26   61 0.13346935 0.42717367 0.19695176 0.23521367  0.167710243
#>    97   62 0.10549624 0.42796597 0.01738048 0.49252307 -0.004011580
#>    40   38 0.84457203 0.16187190 0.60552041 0.01516878  0.779344621
#>    76   12 0.22994046 0.23636208 0.20792567 0.22444154  0.269083281
#>    39   62 0.05966839 0.48888503 0.01738048 0.49252307  0.032506697
#>    93   15 0.57504499 0.10330154 0.45798725 0.21874199  0.479147070
#>    57   61 0.24659482 0.26874459 0.19695176 0.23521367  0.212399320
#>    61   12 0.19695176 0.23521367 0.20792567 0.22444154  0.213269495
#>    41   56 0.42411968 0.01316131 0.60486343 0.00000000  0.514849568
#>    12   87 0.20792567 0.22444154 0.12263840 0.20012488  0.186442080
#>    74   76 0.40337241 0.06109896 0.22994046 0.23636208  0.301578340
#>    86   87 0.03776286 0.25471100 0.12263840 0.20012488  0.123443805
#>    68   32 0.19948783 0.04043276 0.20733439 0.05154818  0.183832643
#>    41   32 0.42411968 0.01316131 0.20733439 0.05154818  0.305393824
#>    32   87 0.20733439 0.05154818 0.12263840 0.20012488  0.176107304
#>     3   50 0.13531639 0.04567593 0.06768192 0.02652085  0.152205342
#>    68   64 0.19948783 0.04043276 0.16260492 0.04677306  0.155127886
#>    64   87 0.16260492 0.04677306 0.12263840 0.20012488  0.145822316
#>    50   41 0.06768192 0.02652085 0.42411968 0.01316131  0.216845619
#>     3   87 0.13531639 0.04567593 0.12263840 0.20012488  0.078098373
#>     1   91 1.00000000 1.00000000 0.91338248 0.83220498  0.915768680
#>    71   55 0.95682906 0.72388305 0.93790499 0.72394215  0.947509154
#>    19   60 0.90621106 0.79076398 0.87089760 0.84224327  0.886990791
#>    60   84 0.87089760 0.84224327 0.81369490 0.75866931  0.797743395
#>    71   19 0.95682906 0.72388305 0.90621106 0.79076398  0.985721924
#>    55   90 0.93790499 0.72394215 0.82042990 0.67676730  0.879913253
#>    19   70 0.90621106 0.79076398 0.81812899 0.71239631  0.871691939
#>    27   65 0.67907927 0.71784563 0.77936952 0.90265472  0.723634585
#>    25   84 0.71172367 0.68698470 0.81369490 0.75866931  0.757698703
#>    25   70 0.71172367 0.68698470 0.81812899 0.71239631  0.766953402
#>    27   10 0.67907927 0.71784563 0.62297677 0.74842619  0.698262485
#>    27   22 0.67907927 0.71784563 0.59991347 0.78680052  0.702556984
#>     5   96 0.57584425 0.69473807 0.63303302 0.72117683  0.615053051
#>    55   95 0.93790499 0.72394215 0.86109722 0.57878841  0.898661989
#>    95   25 0.86109722 0.57878841 0.71172367 0.68698470  0.757753664
#>    51   65 0.55249653 0.98736727 0.77936952 0.90265472  0.650896972
#>    11   71 0.99584295 0.53866786 0.95682906 0.72388305  1.605754646
#>    11   95 0.99584295 0.53866786 0.86109722 0.57878841  0.948313530
#>    80   25 0.71354667 0.43656396 0.71172367 0.68698470  0.732166793
#>    94   27 0.62714701 0.64644988 0.67907927 0.71784563  0.666817550
#>    10    5 0.62297677 0.74842619 0.57584425 0.69473807  0.573175989
#>    49   22 0.46897473 0.87862548 0.59991347 0.78680052  0.539714875
#>    59   67 0.42029734 0.92314035 0.45600321 0.91906964  0.434592751
#>    28    5 0.52104626 0.56325761 0.57584425 0.69473807  0.558427719
#>    28   82 0.52104626 0.56325761 0.50574640 0.68177240  0.552032418
#>    92   49 0.31205899 0.75830901 0.46897473 0.87862548  0.400096752
#>     4   28 0.49161032 0.66809140 0.52104626 0.56325761  0.523481687
#>    94   69 0.62714701 0.64644988 0.65064580 0.41894810  0.631288441
#>     4    9 0.49161032 0.66809140 0.48053435 0.65840022  0.468556233
#>    47   82 0.44331640 0.69443060 0.50574640 0.68177240  0.476506273
#>    80   13 0.71354667 0.43656396 0.84624093 0.50671631  0.808726496
#>    80   94 0.71354667 0.43656396 0.62714701 0.64644988  0.719706381
#>    72   44 0.35138810 0.49923270 0.41985445 0.70782967  0.377389833
#>    72   47 0.35138810 0.49923270 0.44331640 0.69443060  0.388786385
#>    44   45 0.41985445 0.70782967 0.50529775 0.76746954  0.457002147
#>    98   31 0.96257273 0.38659634 0.94379629 0.43323590  1.047186034
#>    48   43 0.90574145 0.41199935 0.71114800 0.20604155  0.796370472
#>    59   92 0.42029734 0.92314035 0.31205899 0.75830901  0.342705921
#>    33   42 0.34345086 0.95865233 0.44289973 0.94738632  0.395026055
#>    48   30 0.90574145 0.41199935 0.98485155 0.30473233  0.855095114
#>    72    9 0.35138810 0.49923270 0.48053435 0.65840022  0.401019373
#>    30   98 0.98485155 0.30473233 0.96257273 0.38659634  1.214421540
#>    33   34 0.34345086 0.95865233 0.20157168 0.94781712  0.275383021
#>    35   51 0.26775950 0.99579311 0.55249653 0.98736727  0.416383833
#>    35    1 0.26775950 0.99579311 1.00000000 1.00000000  0.597531534
#>    81   69 0.47200612 0.37442518 0.65064580 0.41894810  0.547290699
#>    28   81 0.52104626 0.56325761 0.47200612 0.37442518  0.482082362
#>    80   58 0.71354667 0.43656396 0.60409908 0.30807945  0.709776731
#>    66   81 0.52417624 0.32563156 0.47200612 0.37442518  0.530209224
#>    34   73 0.20157168 0.94781712 0.16138271 0.93557897  0.180884928
#>    20   72 0.30651168 0.44614346 0.35138810 0.49923270  0.353781187
#>    66   89 0.52417624 0.32563156 0.57696116 0.34830324  0.545557772
#>    20  100 0.30651168 0.44614346 0.33241176 0.49514781  0.351228127
#>    44    2 0.41985445 0.70782967 0.33115035 0.50051474  0.317264639
#>    34   52 0.20157168 0.94781712 0.14511855 0.80699219  0.204908528
#>     6   35 0.16502639 0.95756678 0.26775950 0.99579311  0.200845667
#>    93   85 0.57504499 0.10330154 0.52070102 0.28728793  0.585178990
#>    85   21 0.52070102 0.28728793 0.54552234 0.31722889  0.568592594
#>    20   54 0.30651168 0.44614346 0.38460384 0.48796418  0.383310458
#>    48   40 0.90574145 0.41199935 0.84457203 0.16187190  0.816684700
#>    92   78 0.31205899 0.75830901 0.16944491 0.64335481  0.214299132
#>    81   85 0.47200612 0.37442518 0.52070102 0.28728793  0.457117174
#>    83   73 0.11962555 0.91535497 0.16138271 0.93557897  0.167626114
#>    46   78 0.20083599 0.45829540 0.16944491 0.64335481  0.239847956
#>    15   37 0.45798725 0.21874199 0.39912401 0.24399954  0.449257291
#>    15   77 0.45798725 0.21874199 0.39075448 0.29857242  0.454990924
#>    81   77 0.47200612 0.37442518 0.39075448 0.29857242  0.380774642
#>    53   77 0.28048834 0.36516186 0.39075448 0.29857242  0.342877681
#>    52   79 0.14511855 0.80699219 0.03768689 0.89191356  0.093424850
#>    53   20 0.28048834 0.36516186 0.30651168 0.44614346  0.366092811
#>    93   58 0.57504499 0.10330154 0.60409908 0.30807945  0.606384481
#>    99   29 0.38012586 0.28949722 0.28492304 0.28644871  0.331225478
#>    46    2 0.20083599 0.45829540 0.33115035 0.50051474  0.259208219
#>    20   16 0.30651168 0.44614346 0.21383728 0.43483484  0.256133259
#>    14   52 0.02630406 0.76670757 0.14511855 0.80699219  0.074331293
#>    79    6 0.03768689 0.89191356 0.16502639 0.95756678  0.036792612
#>    78    7 0.16944491 0.64335481 0.11980498 0.51739171  0.170545562
#>    57   26 0.24659482 0.26874459 0.13346935 0.42717367  0.182109000
#>    26   46 0.13346935 0.42717367 0.20083599 0.45829540  0.173934509
#>    17   78 0.02046862 0.65406897 0.16944491 0.64335481  0.095693314
#>    57   16 0.24659482 0.26874459 0.21383728 0.43483484  0.188081571
#>    78   63 0.16944491 0.64335481 0.16290926 0.47673770  0.199524738
#>    38   93 0.60552041 0.01516878 0.57504499 0.10330154  0.689293967
#>    93   41 0.57504499 0.10330154 0.42411968 0.01316131  0.491966123
#>    17   36 0.02046862 0.65406897 0.09859034 0.58362591  0.083106800
#>    14   23 0.02630406 0.76670757 0.09770875 0.75226346  0.063285743
#>    17   75 0.02046862 0.65406897 0.00000000 0.82756695 -0.164979608
#>    97   63 0.10549624 0.42796597 0.16290926 0.47673770  0.120637028
#>    17    8 0.02046862 0.65406897 0.05962720 0.72899873  0.095935913
#>    97    7 0.10549624 0.42796597 0.11980498 0.51739171  0.117609173
#>    18   88 0.05636180 0.58763117 0.01670961 0.56077328  0.028757235
#>    86   97 0.03776286 0.25471100 0.10549624 0.42796597  0.116543055
#>     7   39 0.11980498 0.51739171 0.05966839 0.48888503  0.072853270
#>    38   43 0.60552041 0.01516878 0.71114800 0.20604155  0.738830366
#>    29   76 0.28492304 0.28644871 0.22994046 0.23636208  0.289496596
#>    18   24 0.05636180 0.58763117 0.00108610 0.58600821  0.027891705
#>    88   62 0.01670961 0.56077328 0.01738048 0.49252307 -0.067949690
#>    15   74 0.45798725 0.21874199 0.40337241 0.06109896  0.394668635
#>    86   62 0.03776286 0.25471100 0.01738048 0.49252307 -0.004011580
#>    56   38 0.60486343 0.00000000 0.60552041 0.01516878  0.779344621
#>    32   12 0.20733439 0.05154818 0.20792567 0.22444154  0.269083281
#>    88   39 0.01670961 0.56077328 0.05966839 0.48888503  0.041637514
#>    76   61 0.22994046 0.23636208 0.19695176 0.23521367  0.212399320
#>    32   76 0.20733439 0.05154818 0.22994046 0.23636208  0.301578340
#>    87   61 0.12263840 0.20012488 0.19695176 0.23521367  0.169768978
#>    86   26 0.03776286 0.25471100 0.13346935 0.42717367  0.126344604
#>    86   24 0.03776286 0.25471100 0.00108610 0.58600821 -1.878315175
#>    32   74 0.20733439 0.05154818 0.40337241 0.06109896  0.308507487
#>    64   32 0.16260492 0.04677306 0.20733439 0.05154818  0.183832643
#>    50   86 0.06768192 0.02652085 0.03776286 0.25471100  0.021765104
#>    68   50 0.19948783 0.04043276 0.06768192 0.02652085  0.152205342
#>     3   64 0.13531639 0.04567593 0.16260492 0.04677306  0.155127886
#>    68   41 0.19948783 0.04043276 0.42411968 0.01316131  0.216845619
#>    50   87 0.06768192 0.02652085 0.12263840 0.20012488  0.078098373
#>    65   60 0.77936952 0.90265472 0.87089760 0.84224327  0.884681054
#>    91   71 0.91338248 0.83220498 0.95682906 0.72388305  1.097758611
#>    11    1 0.99584295 0.53866786 1.00000000 1.00000000  1.605754646
#>    28   94 0.52104626 0.56325761 0.62714701 0.64644988  0.558427719
#>    92   44 0.31205899 0.75830901 0.41985445 0.70782967  0.400096752
#>     4   47 0.49161032 0.66809140 0.44331640 0.69443060  0.468556233
#>    80   48 0.71354667 0.43656396 0.90574145 0.41199935  0.808726496
#>    59   33 0.42029734 0.92314035 0.34345086 0.95865233  0.342705921
#>    54   28 0.38460384 0.48796418 0.52104626 0.56325761  0.423872446
#>    30   11 0.98485155 0.30473233 0.99584295 0.53866786  1.214421540
#>    33   35 0.34345086 0.95865233 0.26775950 0.99579311  0.275383021
#>    34    6 0.20157168 0.94781712 0.16502639 0.95756678  0.180884928
#>    66   21 0.52417624 0.32563156 0.54552234 0.31722889  0.545557772
#>    20   81 0.30651168 0.44614346 0.47200612 0.37442518  0.383310458
#>    83   52 0.11962555 0.91535497 0.14511855 0.80699219  0.167626114
#>    53   99 0.28048834 0.36516186 0.38012586 0.28949722  0.342877681
#>    46   20 0.20083599 0.45829540 0.30651168 0.44614346  0.259208219
#>    57   53 0.24659482 0.26874459 0.28048834 0.36516186  0.188081571
#>    93   74 0.57504499 0.10330154 0.40337241 0.06109896  0.491966123
#>    18    7 0.05636180 0.58763117 0.11980498 0.51739171  0.073046822
#>    17   18 0.02046862 0.65406897 0.05636180 0.58763117  0.083106800
#>    14    8 0.02630406 0.76670757 0.05962720 0.72899873  0.063285743
#>    56   40 0.60486343 0.00000000 0.84457203 0.16187190  0.779344621
#>    61   86 0.19695176 0.23521367 0.03776286 0.25471100  0.123443805
#>    68    3 0.19948783 0.04043276 0.13531639 0.04567593  0.152205342
#>    50   56 0.06768192 0.02652085 0.60486343 0.00000000 -0.040009559
#>            my1           mx2         my2 bp1 bp2
#>   9.372271e-01   0.886990791  0.81543108   0   0
#>   9.626681e-01   0.915768680  0.93722713   0   0
#>   7.694247e-01   0.876330551  0.73566429   0   0
#>   7.926227e-01   0.871691939  0.74087790   0   0
#>   8.309508e-01   0.723634585  0.81328359   0   0
#>   8.432821e-01   1.605754646  0.76385677   0   0
#>   7.356643e-01   0.879913253  0.69849751   0   0
#>   7.983455e-01   0.886990791  0.81543108   0   0
#>   6.984975e-01   0.766953402  0.69120261   0   0
#>   7.408779e-01   0.757698703  0.72995455   0   0
#>   6.261986e-01   0.948313530  0.62537278   0   0
#>   7.519537e-01   0.723634585  0.81328359   0   0
#>   7.458571e-01   0.597271868  0.72345963   0   0
#>   8.197913e-01   0.573175989  0.74461328   0   0
#>   8.247218e-01   0.650896972  0.90474218   0   0
#>   6.819819e-01   0.657962130  0.74585709   0   0
#>   6.849978e-01   0.653341094  0.68198194   0   0
#>   6.518093e-01   0.757753664  0.59332360   0   0
#>   5.933236e-01   0.766953402  0.69120261   0   0
#>   5.665248e-01   0.919662892  0.52914899   0   0
#>   1.223107e+00   0.884681054  0.96266812   0   0
#>   5.619165e-01   0.738342362  0.56652476   0   0
#>   6.721794e-01   0.742233553  0.75195370   0   0
#>   7.446133e-01   0.558464305  0.74845957   0   0
#>   8.860547e-01   0.650896972  0.90474218   0   0
#>   5.291490e-01   0.924064745  0.50853885   0   0
#>   7.484596e-01   0.534041308  0.72477028   0   0
#>   8.402289e-01   0.558464305  0.74845957   0   0
#>   9.198226e-01   0.571851668  0.88605470   0   0
#>   9.283823e-01   0.501388170  0.95726156   0   0
#>   8.899005e-01   0.527887355  0.91982257   0   0
#>   9.572616e-01   0.527887355  0.91982257   0   0
#>   6.248374e-01   0.615053051  0.68499777   0   0
#>   6.275028e-01   0.534041308  0.72477028   0   0
#>   8.006517e-01   0.539714875  0.84022891   0   0
#>   8.059732e-01   0.418600329  0.80065167   0   0
#>   6.204910e-01   0.423872446  0.57807714   0   0
#>   7.244974e-01   0.534041308  0.72477028   0   0
#>   6.266476e-01   0.476506273  0.69784152   0   0
#>   5.319132e-01   0.546582572  0.45584166   0   0
#>   6.832648e-01   0.388786385  0.60086574   0   0
#>   4.725094e-01   0.924064745  0.50853885   0   0
#>   4.171026e-01   0.896926717  0.47250938   0   0
#>   5.618258e-01   0.666817550  0.67217936   0   0
#>   6.062329e-01   0.457002147  0.74563515   0   0
#>   7.456351e-01   0.481910964  0.72449736   0   0
#>   4.477598e-01   0.924064745  0.50853885   0   0
#>   4.029051e-01   0.896926717  0.47250938   0   0
#>   3.204285e-01   0.718794634  0.32123567   0   0
#>   8.561380e-01   0.279166136  0.86609406   0   0
#>   9.693566e-01   0.438979939  0.92838230   0   0
#>   1.182603e+00   0.501388170  0.95726156   0   0
#>   2.918418e-01   4.418815167 -3.20749604   0   1
#>   5.340853e-01   0.709776731  0.32891743   0   0
#>   5.909401e-01   0.423872446  0.57807714   0   0
#>   8.311060e-01   0.434592751  0.88990050   0   0
#>   4.111720e-01   1.047186034  0.44775982   0   0
#>   3.259810e-01   0.935769481  0.40290507   0   0
#>   9.156312e-01   0.200845667  1.01846319   0   0
#>   1.202985e+00   0.419183546  1.18260301   0   0
#>   7.324567e+00   0.769771796  1.22310689   0   0
#>   3.594992e-01   0.575671761  0.31816681   0   0
#>   4.530004e-01   0.636934614  0.35949922   0   0
#>   4.725925e-01   0.383310458  0.39655791   0   0
#>   3.289174e-01   0.636934614  0.35949922   0   0
#>   3.843690e-01   0.547290699  0.45300043   0   0
#>   9.436430e-01   0.127409122  0.95250466   0   0
#>   8.660941e-01   0.247546577  0.84765918   0   0
#>   4.516981e-01   0.401019373  0.59094007   0   0
#>   6.156806e-01   0.341281287  0.50006416   0   0
#>   3.486340e-01   0.575671761  0.31816681   0   0
#>   4.538563e-01   0.341281287  0.50006416   0   0
#>   6.290905e-01   0.299631766  0.62778464   0   0
#>   3.118986e-01   0.528731796  0.30588937   0   0
#>   8.476592e-01   0.214299132  0.73364971   0   0
#>   8.647517e-01   0.167626114  0.86946731   0   0
#>   6.277846e-01   0.239847956  0.56010498   0   0
#>   3.212357e-01   0.606384481  0.20330513   0   0
#>   2.063138e-01   0.568592594  0.27284442   0   0
#>   2.728444e-01   0.575671761  0.31816681   0   0
#>   3.012351e-01   0.738830366  0.06605919   0   0
#>   7.336497e-01   0.180320693  0.72859847   0   0
#>   3.089301e-01   0.454990924  0.28444525   0   0
#>   4.874589e-01   0.341281287  0.50006416   0   0
#>   7.285985e-01   0.113077619  0.68430668   0   0
#>   2.796164e-01   0.403714975  0.27263186   0   0
#>   3.907064e-01   0.366092811  0.38232504   0   0
#>   3.438829e-01   0.403714975  0.27263186   0   0
#>   5.577102e-01   0.168453193  0.43991940   0   0
#>   8.520110e-01   0.036792612  1.04996710   0   0
#>   1.963127e-01   0.454990924  0.28444525   0   0
#>   3.285389e-01   0.244485598  0.32365196   0   0
#>   5.003475e-01   0.287650048  0.48745894   0   0
#>   4.153771e-01   0.188081571  0.34347964   0   0
#>   4.736071e-01   0.173934509  0.42805445   0   0
#>   2.726319e-01   0.333948533  0.24349998   0   0
#>   8.274363e-01 -11.497491828  7.60469249   0   1
#>   8.204138e-01   0.180320693  0.72859847   0   0
#>   1.049967e+00   0.127409122  0.95250466   0   0
#>   5.701584e-01   0.073046822  0.53892979   0   0
#>   2.435000e-01   0.316501328  0.19656178   0   0
#>   3.423017e-01   0.173934509  0.42805445   0   0
#>   6.589533e-01   0.170545562  0.57015838   0   0
#>   5.587382e-01   0.117609173  0.47188544   0   0
#>   9.347226e-02   0.606384481  0.20330513   0   0
#>   7.098355e-02   0.308507487 -0.00841678   0   0
#>   6.449948e-01   0.073046822  0.53892979   0   0
#>   7.658099e-01   0.113077619  0.68430668   0   0
#>   1.965618e-01   0.318574876  0.15062892   0   0
#>   7.201469e-01   0.073395117  0.82317495   0   0
#>   8.231750e-01   0.073999283  0.82743627   0   0
#>   4.683211e-01   0.168453193  0.43991940   0   0
#>   6.623266e-01   0.113077619  0.68430668   0   0
#>   5.856861e-01  -0.067949690  0.52581271   0   0
#>   7.115049e-01  -0.164979608  0.72014690   0   0
#>   3.306859e-02   0.514849568  0.01149729   0   0
#>   3.236520e-01   0.289496596  0.22620622   0   0
#>   3.237797e-01   0.120637028  0.46832111   0   0
#>   4.741462e-01   0.032506697  0.42075490   0   0
#>   5.387550e-01   0.060818401  0.53835197   0   0
#>   2.262062e-01   0.212399320  0.26585744   0   0
#>   6.151647e-01  -0.289261928  0.70548455   0   0
#>   5.258127e-01  -1.878315175  0.21026746   0   0
#>   7.054846e-01  -4.428958069  0.68687159   0   1
#>   1.523964e-01   0.318574876  0.15062892   0   0
#>   3.320203e-01   0.126344604  0.31834045   0   0
#>   3.709101e-01   0.032506697  0.42075490   0   0
#>   4.151386e-05   0.738830366  0.06605919   0   0
#>   1.377847e-01   0.213269495  0.24086126   0   0
#>   4.207549e-01   0.041637514  0.52688991   0   0
#>   1.231292e-01   0.551319492  0.19631272   0   0
#>   2.658574e-01   0.167710243  0.33202033   0   0
#>   2.408613e-01   0.169768978  0.19654590   0   0
#>   1.149729e-02  -0.040009559 -7.60835938   0   0
#>   1.380673e-01   0.169768978  0.19654590   0   0
#>   1.338100e-01   0.318574876  0.15062892   0   0
#>   2.946564e-01   0.021765104  0.13655697   0   0
#>   5.981124e-02   0.305393824 -0.02600082   0   0
#>  -2.600082e-02   0.308507487 -0.00841678   0   0
#>   1.321760e-01   0.186442080  0.13806732   0   0
#>  -1.429395e-01   0.078098373  0.11872399   0   0
#>  -1.071704e-01   0.183832643  0.05981124   0   0
#>   1.242831e-01   0.176107304  0.13217599   0   0
#>  -7.553625e-01  -0.040009559 -7.60835938   0   0
#>   1.187240e-01   0.145822316  0.12428313   0   0
#>   9.372271e-01   1.097758611  0.84328205   0   0
#>   7.694247e-01   0.947061870  0.62619861   0   0
#>   8.154311e-01   0.853741201  0.79262273   0   0
#>   8.309508e-01   0.853741201  0.79262273   0   0
#>   7.983455e-01   0.947509154  0.76942466   0   0
#>   6.984975e-01   0.898661989  0.65180930   0   0
#>   7.408779e-01   0.876330551  0.73566429   0   0
#>   8.132836e-01   0.702556984  0.82472176   0   0
#>   7.299545e-01   0.742233553  0.75195370   0   0
#>   6.912026e-01   0.757698703  0.72995455   0   0
#>   8.197913e-01   0.657962130  0.74585709   0   0
#>   8.247218e-01   0.698262485  0.81979133   0   0
#>   6.849978e-01   0.597271868  0.72345963   0   0
#>   6.518093e-01   0.947061870  0.62619861   0   0
#>   5.933236e-01   0.738342362  0.56652476   0   0
#>   9.047422e-01   0.769771796  1.22310689   0   0
#>   7.638568e-01   0.948313530  0.62537278   0   0
#>   6.253728e-01   0.919662892  0.52914899   0   0
#>   5.619165e-01   0.719706381  0.56182581   0   0
#>   6.721794e-01   0.653341094  0.68198194   0   0
#>   7.446133e-01   0.597271868  0.72345963   0   0
#>   8.402289e-01   0.571851668  0.88605470   0   0
#>   8.899005e-01   0.438979939  0.92838230   0   0
#>   6.248374e-01   0.552032418  0.62750279   0   0
#>   6.275028e-01   0.545408205  0.62664763   0   0
#>   8.059732e-01   0.380825955  0.83110605   0   0
#>   6.204910e-01   0.545408205  0.62664763   0   0
#>   5.319132e-01   0.652317911  0.53408531   0   0
#>   6.832648e-01   0.523481687  0.62049095   0   0
#>   6.978415e-01   0.481910964  0.72449736   0   0
#>   4.171026e-01   0.732166793  0.56191651   0   0
#>   5.618258e-01   0.652317911  0.53408531   0   0
#>   6.062329e-01   0.348605433  0.61568064   0   0
#>   6.008657e-01   0.377389833  0.60623294   0   0
#>   7.456351e-01   0.418600329  0.80065167   0   0
#>   4.477598e-01   0.935769481  0.40290507   0   0
#>   3.204285e-01   0.816684700  0.30123514   0   0
#>   8.561380e-01   0.380825955  0.83110605   0   0
#>   9.693566e-01   0.419183546  1.18260301   0   0
#>   2.918418e-01   0.901385184  0.32598097   0   0
#>   5.909401e-01   0.388786385  0.60086574   0   0
#>   4.111720e-01   0.901385184  0.32598097   0   0
#>   9.156312e-01   0.279166136  0.86609406   0   0
#>   1.202985e+00   0.597531534  7.32456750   0   0
#>   7.324567e+00   0.589685687  8.69019395   0   1
#>   4.530004e-01   0.546582572  0.45584166   0   0
#>   4.725925e-01   0.546582572  0.45584166   0   0
#>   3.289174e-01   0.718794634  0.32123567   0   0
#>   3.843690e-01   0.462429219  0.31189861   0   0
#>   9.436430e-01   0.204908528  0.86475169   0   0
#>   4.516981e-01   0.351228127  0.45385626   0   0
#>   3.486340e-01   0.530209224  0.38436897   0   0
#>   4.538563e-01   0.287650048  0.48745894   0   0
#>   6.290905e-01   0.348605433  0.61568064   0   0
#>   8.647517e-01   0.247546577  0.84765918   0   0
#>   1.018463e+00  -2.980613158  9.56862257   0   1
#>   2.063138e-01   0.551319492  0.19631272   0   0
#>   2.728444e-01   0.528731796  0.30588937   0   0
#>   3.965579e-01   0.353781187  0.45169815   0   0
#>   3.012351e-01   0.855095114  0.29184175   0   0
#>   7.336497e-01   0.299631766  0.62778464   0   0
#>   3.089301e-01   0.462429219  0.31189861   0   0
#>   8.694673e-01   0.127409122  0.95250466   0   0
#>   5.601050e-01   0.225730329  0.55771025   0   0
#>   2.796164e-01   0.394668635  0.15239642   0   0
#>   2.844453e-01   0.449257291  0.27961642   0   0
#>   3.907064e-01   0.457117174  0.30893009   0   0
#>   3.438829e-01   0.366092811  0.38232504   0   0
#>   8.520110e-01   0.073999283  0.82743627   0   0
#>   3.823250e-01   0.263238789  0.41537707   0   0
#>   2.033051e-01   0.585178990  0.20631378   0   0
#>   3.285389e-01   0.333948533  0.24349998   0   0
#>   5.003475e-01   0.239847956  0.56010498   0   0
#>   4.736071e-01   0.263238789  0.41537707   0   0
#>   8.204138e-01   0.073395117  0.82317495   0   0
#>   1.049967e+00  -3.161794582  7.25387465   0   1
#>   5.701584e-01   0.199524738  0.55873818   0   0
#>   3.423017e-01   0.167710243  0.33202033   0   0
#>   4.280544e-01   0.168453193  0.43991940   0   0
#>   6.589533e-01   0.095935913  0.66232657   0   0
#>   3.434796e-01   0.182109000  0.34230169   0   0
#>   5.587382e-01   0.225730329  0.55771025   0   0
#>   9.347226e-02   0.514610850  0.03306859   0   0
#>   7.098355e-02   0.514610850  0.03306859   0   0
#>   6.449948e-01   0.095693314  0.65895332   0   0
#>   7.658099e-01   0.074331293  0.82041382   0   0
#>   7.201469e-01  -0.289261928  0.70548455   0   0
#>   4.683211e-01   0.117609173  0.47188544   0   0
#>   6.623266e-01   0.001833491  0.71150485   0   0
#>   4.718854e-01   0.103479896  0.47414622   0   0
#>   5.856861e-01   0.060818401  0.53835197   0   0
#>   3.237797e-01  -0.004011580  0.37091010   0   0
#>   5.387550e-01   0.103479896  0.47414622   0   0
#>   6.605919e-02   0.689293967  0.09347226   0   0
#>   2.262062e-01   0.316501328  0.19656178   0   0
#>   6.151647e-01   0.028757235  0.58568613   0   0
#>   5.258127e-01   0.041637514  0.52688991   0   0
#>   1.523964e-01   0.479147070  0.12312918   0   0
#>   3.709101e-01  -1.878315175  0.21026746   0   0
#>   4.151386e-05   0.514849568  0.01149729   0   0
#>   1.377847e-01   0.186442080  0.13806732   0   0
#>   5.268899e-01   0.060818401  0.53835197   0   0
#>   2.658574e-01   0.213269495  0.24086126   0   0
#>   1.338100e-01   0.269083281  0.13778470   0   0
#>   1.965459e-01   0.123443805  0.29465643   0   0
#>   3.183404e-01   0.116543055  0.32377972   0   0
#>   2.102675e-01  -4.860207051 -0.11984740   0   1
#>  -8.416780e-03   0.301578340  0.13380996   0   0
#>   5.981124e-02   0.176107304  0.13217599   0   0
#>   1.365570e-01  -4.286481828 -0.42831718   0   1
#>  -1.429395e-01   0.216845619 -0.75536253   0   0
#>  -1.071704e-01   0.145822316  0.12428313   0   0
#>  -7.553625e-01   0.305393824 -0.02600082   0   0
#>   1.187240e-01   0.021765104  0.13655697   0   0
#>   9.626681e-01   0.797743395  0.83095075   0   0
#>   8.432821e-01   0.985721924  0.79834553   0   0
#>   7.638568e-01   3.773214322  0.74432587   0   1
#>   6.248374e-01   0.631288441  0.53191316   0   0
#>   8.059732e-01   0.317264639  0.62909047   0   0
#>   6.832648e-01   0.476506273  0.69784152   0   0
#>   4.171026e-01   0.796370472  0.32042846   0   0
#>   8.561380e-01   0.395026055  0.96935663   0   0
#>   5.780771e-01   0.482082362  0.47259249   0   0
#>   4.111720e-01   5.479687725  0.21076948   0   1
#>   9.156312e-01   0.416383833  1.20298501   0   0
#>   9.436430e-01   0.200845667  1.01846319   0   0
#>   3.486340e-01   0.528731796  0.30588937   0   0
#>   3.965579e-01   0.380774642  0.39070635   0   0
#>   8.694673e-01   0.093424850  0.85201101   0   0
#>   3.438829e-01   0.331225478  0.32853889   0   0
#>   5.003475e-01   0.256133259  0.47360707   0   0
#>   3.434796e-01   0.244485598  0.32365196   0   0
#>   7.098355e-02   0.479147070  0.12312918   0   0
#>   5.389298e-01   0.072853270  0.53875497   0   0
#>   6.449948e-01   0.027891705  0.61516472   0   0
#>   7.658099e-01   0.001833491  0.71150485   0   0
#>   4.151386e-05   2.714159618 -2.86513648   0   1
#>   2.946564e-01   0.126344604  0.31834045   0   0
#>  -1.429395e-01   0.155127886 -0.10717042   0   0
#>  -7.608359e+00  -0.131692409 -9.46540116   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 1.00000000 1.00000000
#>   [2,] 0.33115035 0.50051474
#>   [3,] 0.13531639 0.04567593
#>   [4,] 0.49161032 0.66809140
#>   [5,] 0.57584425 0.69473807
#>   [6,] 0.16502639 0.95756678
#>   [7,] 0.11980498 0.51739171
#>   [8,] 0.05962720 0.72899873
#>   [9,] 0.48053435 0.65840022
#>  [10,] 0.62297677 0.74842619
#>  [11,] 0.99584295 0.53866786
#>  [12,] 0.20792567 0.22444154
#>  [13,] 0.84624093 0.50671631
#>  [14,] 0.02630406 0.76670757
#>  [15,] 0.45798725 0.21874199
#>  [16,] 0.21383728 0.43483484
#>  [17,] 0.02046862 0.65406897
#>  [18,] 0.05636180 0.58763117
#>  [19,] 0.90621106 0.79076398
#>  [20,] 0.30651168 0.44614346
#>  [21,] 0.54552234 0.31722889
#>  [22,] 0.59991347 0.78680052
#>  [23,] 0.09770875 0.75226346
#>  [24,] 0.00108610 0.58600821
#>  [25,] 0.71172367 0.68698470
#>  [26,] 0.13346935 0.42717367
#>  [27,] 0.67907927 0.71784563
#>  [28,] 0.52104626 0.56325761
#>  [29,] 0.28492304 0.28644871
#>  [30,] 0.98485155 0.30473233
#>  [31,] 0.94379629 0.43323590
#>  [32,] 0.20733439 0.05154818
#>  [33,] 0.34345086 0.95865233
#>  [34,] 0.20157168 0.94781712
#>  [35,] 0.26775950 0.99579311
#>  [36,] 0.09859034 0.58362591
#>  [37,] 0.39912401 0.24399954
#>  [38,] 0.60552041 0.01516878
#>  [39,] 0.05966839 0.48888503
#>  [40,] 0.84457203 0.16187190
#>  [41,] 0.42411968 0.01316131
#>  [42,] 0.44289973 0.94738632
#>  [43,] 0.71114800 0.20604155
#>  [44,] 0.41985445 0.70782967
#>  [45,] 0.50529775 0.76746954
#>  [46,] 0.20083599 0.45829540
#>  [47,] 0.44331640 0.69443060
#>  [48,] 0.90574145 0.41199935
#>  [49,] 0.46897473 0.87862548
#>  [50,] 0.06768192 0.02652085
#>  [51,] 0.55249653 0.98736727
#>  [52,] 0.14511855 0.80699219
#>  [53,] 0.28048834 0.36516186
#>  [54,] 0.38460384 0.48796418
#>  [55,] 0.93790499 0.72394215
#>  [56,] 0.60486343 0.00000000
#>  [57,] 0.24659482 0.26874459
#>  [58,] 0.60409908 0.30807945
#>  [59,] 0.42029734 0.92314035
#>  [60,] 0.87089760 0.84224327
#>  [61,] 0.19695176 0.23521367
#>  [62,] 0.01738048 0.49252307
#>  [63,] 0.16290926 0.47673770
#>  [64,] 0.16260492 0.04677306
#>  [65,] 0.77936952 0.90265472
#>  [66,] 0.52417624 0.32563156
#>  [67,] 0.45600321 0.91906964
#>  [68,] 0.19948783 0.04043276
#>  [69,] 0.65064580 0.41894810
#>  [70,] 0.81812899 0.71239631
#>  [71,] 0.95682906 0.72388305
#>  [72,] 0.35138810 0.49923270
#>  [73,] 0.16138271 0.93557897
#>  [74,] 0.40337241 0.06109896
#>  [75,] 0.00000000 0.82756695
#>  [76,] 0.22994046 0.23636208
#>  [77,] 0.39075448 0.29857242
#>  [78,] 0.16944491 0.64335481
#>  [79,] 0.03768689 0.89191356
#>  [80,] 0.71354667 0.43656396
#>  [81,] 0.47200612 0.37442518
#>  [82,] 0.50574640 0.68177240
#>  [83,] 0.11962555 0.91535497
#>  [84,] 0.81369490 0.75866931
#>  [85,] 0.52070102 0.28728793
#>  [86,] 0.03776286 0.25471100
#>  [87,] 0.12263840 0.20012488
#>  [88,] 0.01670961 0.56077328
#>  [89,] 0.57696116 0.34830324
#>  [90,] 0.82042990 0.67676730
#>  [91,] 0.91338248 0.83220498
#>  [92,] 0.31205899 0.75830901
#>  [93,] 0.57504499 0.10330154
#>  [94,] 0.62714701 0.64644988
#>  [95,] 0.86109722 0.57878841
#>  [96,] 0.63303302 0.72117683
#>  [97,] 0.10549624 0.42796597
#>  [98,] 0.96257273 0.38659634
#>  [99,] 0.38012586 0.28949722
#> [100,] 0.33241176 0.49514781
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,60,91), [4,7,2]
#> 2: (65,60,1), [1,26,6]
#> 3: (71,19,55), [8,13,9]
#> 4: (19,91,60), [1,5,9]
#> 5: (60,84,19), [11,4,6]
#> 6: (60,65,84), [12,5,2]
#> 7: (91,71,1), [27,1,9]
#> 8: (19,70,55), [10,3,11]
#> 9: (71,91,19), [4,3,7]
#> 10: (55,70,90), [16,22,8]
#> 11: (19,84,70), [15,8,5]
#> 12: (27,84,65), [6,19,14]
#> 13: (55,95,71), [28,3,22]
#> 14: (25,84,27), [12,30,15]
#> 15: (25,70,84), [11,14,16]
#> 16: (25,90,70), [10,15,23]
#> 17: (27,10,96), [32,20,18]
#> 18: (27,22,10), [31,17,19]
#> 19: (27,65,22), [25,18,12]
#> 20: (94,27,96), [17,21,30]
#> 21: (5,94,96), [20,32,41]
#> 22: (55,90,95), [23,13,10]
#> 23: (95,90,25), [16,24,22]
#> 24: (25,13,95), [34,23,29]
#> 25: (51,22,65), [19,26,33]
#> 26: (51,65,1), [2,74,25]
#> 27: (11,1,71), [7,28,0]
#> 28: (11,71,95), [13,34,27]
#> 29: (80,13,25), [24,55,54]
#> 30: (94,25,27), [14,20,55]
#> 31: (10,22,5), [35,32,18]
#> 32: (5,96,10), [17,31,21]
#> 33: (49,22,51), [25,37,36]
#> 34: (95,13,11), [48,28,24]
#> 35: (22,45,5), [43,31,36]
#> 36: (49,45,22), [35,33,44]
#> 37: (67,49,51), [33,40,39]
#> 38: (59,67,42), [40,63,39]
#> 39: (59,49,67), [37,38,69]
#> 40: (42,67,51), [37,64,38]
#> 41: (28,94,5), [21,42,50]
#> 42: (28,5,82), [43,49,41]
#> 43: (82,5,45), [35,47,42]
#> 44: (44,45,49), [36,45,58]
#> 45: (92,44,49), [44,69,89]
#> 46: (4,9,28), [68,49,51]
#> 47: (47,82,45), [43,58,53]
#> 48: (31,11,13), [34,52,59]
#> 49: (28,82,4), [53,46,42]
#> 50: (94,28,69), [84,66,41]
#> 51: (4,47,9), [57,46,53]
#> 52: (48,31,13), [48,54,60]
#> 53: (47,4,82), [49,47,51]
#> 54: (80,48,13), [52,29,61]
#> 55: (80,25,94), [30,66,29]
#> 56: (72,47,44), [58,83,57]
#> 57: (72,9,47), [51,56,67]
#> 58: (44,47,45), [47,44,56]
#> 59: (98,11,31), [48,60,70]
#> 60: (98,31,48), [52,71,59]
#> 61: (48,80,43), [95,100,54]
#> 62: (59,33,92), [81,69,63]
#> 63: (33,59,42), [38,64,62]
#> 64: (33,42,51), [40,73,63]
#> 65: (48,40,30), [0,71,100]
#> 66: (94,69,80), [78,55,50]
#> 67: (72,54,9), [68,57,82]
#> 68: (54,28,9), [46,67,77]
#> 69: (92,49,59), [39,62,45]
#> 70: (30,11,98), [59,71,0]
#> 71: (30,98,48), [60,65,70]
#> 72: (33,35,34), [93,81,73]
#> 73: (35,33,51), [64,74,72]
#> 74: (35,51,1), [26,0,73]
#> 75: (69,89,58), [85,78,76]
#> 76: (81,89,69), [75,84,79]
#> 77: (28,54,81), [99,84,68]
#> 78: (80,69,58), [75,95,66]
#> 79: (66,89,81), [76,90,86]
#> 80: (34,6,73), [105,92,93]
#> 81: (33,34,92), [91,62,72]
#> 82: (20,54,72), [67,87,99]
#> 83: (44,2,72), [88,56,89]
#> 84: (81,69,28), [50,77,76]
#> 85: (21,58,89), [75,86,97]
#> 86: (66,21,89), [85,79,98]
#> 87: (20,72,100), [88,103,82]
#> 88: (100,72,2), [83,103,87]
#> 89: (44,92,2), [94,83,45]
#> 90: (81,85,66), [98,79,102]
#> 91: (34,52,92), [101,81,92]
#> 92: (34,73,52), [104,91,80]
#> 93: (6,34,35), [72,0,80]
#> 94: (92,78,2), [106,89,101]
#> 95: (80,58,43), [116,61,78]
#> 96: (93,58,85), [97,114,116]
#> 97: (85,58,21), [85,98,96]
#> 98: (21,66,85), [90,97,86]
#> 99: (20,81,54), [77,82,110]
#> 100: (48,43,40), [152,65,61]
#> 101: (92,52,78), [107,94,91]
#> 102: (81,77,85), [109,90,110]
#> 103: (20,100,2), [88,118,87]
#> 104: (83,52,73), [92,105,113]
#> 105: (83,73,6), [80,124,104]
#> 106: (46,2,78), [94,112,118]
#> 107: (52,23,78), [129,101,123]
#> 108: (15,77,37), [121,158,109]
#> 109: (15,85,77), [102,108,114]
#> 110: (81,20,77), [115,102,99]
#> 111: (53,99,77), [121,115,117]
#> 112: (78,63,46), [139,106,132]
#> 113: (52,83,79), [124,122,104]
#> 114: (93,85,15), [109,165,96]
#> 115: (53,77,20), [110,119,111]
#> 116: (93,43,58), [95,96,133]
#> 117: (99,53,29), [148,126,111]
#> 118: (46,20,2), [103,106,120]
#> 119: (20,16,53), [131,115,120]
#> 120: (20,46,16), [128,119,118]
#> 121: (77,99,37), [126,108,111]
#> 122: (52,79,75), [0,141,113]
#> 123: (14,23,52), [107,141,137]
#> 124: (79,83,6), [105,0,113]
#> 125: (78,36,7), [135,132,130]
#> 126: (99,29,37), [138,121,117]
#> 127: (57,16,26), [128,159,131]
#> 128: (26,16,46), [120,139,127]
#> 129: (8,78,23), [107,137,143]
#> 130: (17,36,78), [125,143,136]
#> 131: (57,53,16), [119,127,148]
#> 132: (78,7,63), [144,112,125]
#> 133: (38,43,93), [116,147,152]
#> 134: (93,74,41), [176,147,165]
#> 135: (18,7,36), [125,136,151]
#> 136: (17,18,36), [135,130,155]
#> 137: (14,8,23), [129,123,146]
#> 138: (29,76,37), [168,126,153]
#> 139: (26,46,63), [112,142,128]
#> 140: (17,14,75), [141,157,146]
#> 141: (14,52,75), [122,140,123]
#> 142: (97,26,63), [139,144,149]
#> 143: (17,78,8), [129,146,130]
#> 144: (97,63,7), [132,150,142]
#> 145: (18,24,88), [156,154,155]
#> 146: (8,14,17), [140,143,137]
#> 147: (93,41,38), [169,133,134]
#> 148: (53,57,29), [153,117,131]
#> 149: (86,26,97), [142,160,174]
#> 150: (7,39,97), [163,144,151]
#> 151: (7,18,39), [154,150,135]
#> 152: (38,40,43), [100,133,161]
#> 153: (29,57,76), [166,138,148]
#> 154: (88,39,18), [151,145,164]
#> 155: (18,17,24), [157,145,136]
#> 156: (88,24,62), [175,164,145]
#> 157: (17,75,24), [0,155,140]
#> 158: (15,37,74), [168,165,108]
#> 159: (57,26,61), [174,166,127]
#> 160: (86,97,62), [163,175,149]
#> 161: (56,40,38), [152,169,0]
#> 162: (32,76,12), [167,170,171]
#> 163: (97,39,62), [164,160,150]
#> 164: (88,62,39), [163,154,156]
#> 165: (74,93,15), [114,158,134]
#> 166: (76,57,61), [159,167,153]
#> 167: (76,61,12), [172,162,166]
#> 168: (74,37,76), [138,171,158]
#> 169: (38,41,56), [184,161,147]
#> 170: (32,12,87), [172,180,162]
#> 171: (32,74,76), [168,162,176]
#> 172: (87,12,61), [167,173,170]
#> 173: (61,86,87), [179,172,174]
#> 174: (86,61,26), [159,149,173]
#> 175: (86,62,24), [156,0,160]
#> 176: (32,41,74), [134,171,178]
#> 177: (64,68,32), [178,180,182]
#> 178: (68,41,32), [176,177,185]
#> 179: (50,87,86), [173,0,186]
#> 180: (64,32,87), [170,183,177]
#> 181: (68,3,50), [186,185,182]
#> 182: (3,68,64), [177,183,181]
#> 183: (3,64,87), [180,186,182]
#> 184: (50,56,41), [169,185,0]
#> 185: (68,50,41), [184,178,181]
#> 186: (50,3,87), [183,179,181]
#> boundary nodes:  11 1 35 6 79 75 24 86 50 56 40 30 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.043654695 0.203849491 0.073957163 0.097925003 0.148020376 0.279471465 
#>                                                                         
#> 0.120331191 0.042056933 0.035703229 0.046484958 0.173830492 0.140669621 
#>                                                                         
#> 0.029045747 0.044771699 0.213604000 0.046166590 0.074958403 0.106083424 
#>                                                                         
#> 0.109185345 0.073587348 0.241151225 0.224925733 0.044922756 0.095156831 
#>                                                                         
#> 0.206095579 0.152976037 0.101324608 0.096570293 0.137115527 0.031201536 
#>                                                                         
#> 0.042473405 0.118218138 0.070453689 0.071286858 0.116940172 0.177718910 
#>                                                                         
#> 0.103408566 0.085698320 0.019672285 0.193962045 0.051801190 0.122132792 
#>                                                                         
#> 0.111855323 0.093788497 0.027018488 0.095793399 0.117578710 0.043579380 
#>                                                                         
#> 0.230534892 0.202787808 0.033147174 0.116661626 0.200218495 0.065321037 
#>                                                                         
#> 0.195578902 0.065962575 0.155668414 0.062250361 0.081746696 0.211008630 
#>                                                                         
#> 0.447681740 0.048522367 0.102078997 0.143283852 0.120243311 0.108156901 
#>                                                                         
#> 0.022287666 0.219364471 0.035075140 0.020278319 0.044204237 0.019411019 
#>                                                                         
#> 0.258500225 0.038500796 0.173894113 0.129611280 0.215758972 0.147889160 
#>                                                                         
#> 0.085950707 0.059286984 0.140545123 0.165435678 0.130435585 0.005513178 
#>                                                                         
#> 0.130411548 0.055210942 0.169923693 0.013975933 0.042172918 0.085225838 
#>                                                                         
#> 0.092906203 0.078837978 0.059693385 0.096419333 0.026822223 0.049304859 
#>                                                                         
#> 0.074570695 0.072408029 0.061992521 0.069548764 0.121835104 0.080732262 
#>                                                                         
#> 0.092670984 0.059251390 0.170527236 0.052234734 0.042418054 0.044625688 
#>                                                                         
#> 0.169355848 0.066300586 0.146569829 0.057648090 0.139265253 0.029679881 
#>                                                                         
#> 0.112789655 0.181411847 0.042219528 0.027984331 0.076232035 0.098801488 
#>                                                                         
#> 0.036414205 0.070766867 0.094894563 0.241561182 0.182949908 0.202184713 
#>                                                                         
#> 0.109233714 0.280477236 0.025034978 0.042444108 0.164405027 0.059906223 
#>                                                                         
#> 0.015377435 0.181222305 0.088686065 0.246568046 0.100913339 0.013605928 
#>                                                                         
#> 0.220157696 0.171021770 0.070294657 0.037423903 0.158474296 0.356688027 
#>                                                                         
#> 0.154968416 0.188832640 0.018924160 0.062427225 0.101275644 0.083876339 
#>                                                                         
#> 0.126593299 0.117898011 0.210267770 0.124646764 0.109397637 0.063895704 
#>                                                                         
#> 0.104985716 0.063004477 0.164222538 0.184442111 0.242172599 0.189279557 
#>                                                                         
#> 0.140591863 0.250427369 0.088285398 0.071441509 0.159927414 0.035937158 
#>                                                                         
#> 0.142442724 0.119498288 0.197733664 0.108888005 0.228712168 0.014717198 
#>                                                                         
#> 0.063700351 0.150097030 0.226973567 0.219545747 0.215761511 0.104199197 
#>                                                                         
#> 0.050277262 0.283346478 0.197192572 0.100084971 0.133283986 0.204970856 
#>                                                                         
#> 0.084841391 0.142292311 0.284861669 0.732252583 0.184104383 0.195096443 
#>                                                                         
#> 0.168781053 0.071432055 0.042011022 0.069515186 0.057447827 0.055427797 
#>                                                                         
#> 0.225494785 0.151718880 0.109614530 0.191844364 0.038891636 0.088585317 
#>                                                                         
#> 0.257498425 0.183175449 0.099820306 0.046396881 0.187702911 0.064053299 
#>                                                                         
#> 0.104370217 0.111155165 0.128812952 0.136942332 0.085060175 0.206828757 
#>                                                                         
#> 0.095251615 0.136982863 0.093361819 0.125458096 0.143267903 0.135391378 
#>                                                                         
#> 0.194671894 0.074207994 0.149361067 0.169289781 0.166745237 0.093253062 
#>                                                                         
#> 0.175794512 0.105191385 0.072850955 0.174701210 0.075332178 0.084545039 
#>                                                                         
#> 0.090563250 0.047891992 0.186024453 0.066551032 0.218150404 0.074375767 
#>                                                                         
#> 0.055299525 0.068253512 0.166835559 0.238683938 0.015183004 0.172894365 
#>                                                                         
#> 0.083745912 0.033008681 0.186191328 0.082180886 0.197238699 0.333321205 
#>                                                                         
#> 0.196270532 0.044983637 0.230143201 0.132538060 0.027310569 0.226281239 
#>                                                                         
#> 0.182094960 0.109667371 0.116710088 0.461350868 0.134827013 0.119029507 
#>                                                                         
#> 0.055009593 0.193758235 0.084655087 0.155838490 0.234193599 0.084312621 
#>                                                                         
#> 0.037823464 0.022940377 0.180366076 0.111321096 0.125111043 0.106372094 
#>                                                                         
#> 0.102201084 0.176783857 0.094649980 0.075513585 0.050322837 0.289245099 
#>                                     
#> 0.160378467 0.064385276 0.537835777 
#> 
#> $alpha
#> [1] 0.3379983
#> 
#> attr(,"class")
#> [1] "scree"