Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1          y1          x2           y2          mx1
#>     9   28 0.20764987 0.809207112 0.131447986 0.7749064565  0.185957638
#>    28   37 0.13144799 0.774906457 0.191035414 0.6250357253  0.163320185
#>    37   52 0.19103541 0.625035725 0.282747345 0.5610843758  0.276749498
#>    52   64 0.28274735 0.561084376 0.260158109 0.7505433705  0.320213137
#>    64    9 0.26015811 0.750543371 0.207649868 0.8092071121  0.222848804
#>    80   11 0.14018049 0.416342353 0.171493140 0.3397613781  0.175164303
#>    11   98 0.17149314 0.339761378 0.266056715 0.3366930966  0.219656161
#>    98   21 0.26605672 0.336693097 0.338325627 0.3503164638  0.288828266
#>    21   78 0.33832563 0.350316464 0.382356276 0.4665283320  0.298070664
#>    78    4 0.38235628 0.466528332 0.208137518 0.4807916769  0.293269811
#>     4   80 0.20813752 0.480791677 0.140180488 0.4163423533  0.172964651
#>    39   71 0.85893360 0.944625662 0.632468787 0.9404581777  0.745640560
#>    71    8 0.63246879 0.940458178 0.581638392 0.7980059757  0.653097200
#>     8   26 0.58163839 0.798005976 0.574245801 0.7057757707  0.676769371
#>    26   18 0.57424580 0.705775771 0.794396431 0.7111173080  0.683679414
#>    18   10 0.79439643 0.711117308 0.823828793 0.8241718864  0.776997685
#>    10   39 0.82382879 0.824171886 0.858933605 0.9446256620  0.789598523
#>    78   52 0.38235628 0.466528332 0.282747345 0.5610843758  0.295322870
#>    37   25 0.19103541 0.625035725 0.111733484 0.4715129652  0.153078091
#>    25   80 0.11173348 0.471512965 0.140180488 0.4163423533  0.161283885
#>    91   65 0.28812285 0.783283496 0.453680392 0.7744574186  0.371248982
#>    65   58 0.45368039 0.774457419 0.442988108 0.9078684974  0.406810132
#>    58   69 0.44298811 0.907868497 0.274594028 0.9733896914  0.359882693
#>    69   75 0.27459403 0.973389691 0.285003567 0.8700315128  0.324330086
#>    75   91 0.28500357 0.870031513 0.288122852 0.7832834957  0.297986315
#>    62   84 0.97585554 0.721226759 0.909217141 0.8175022692  0.924239264
#>    84   47 0.90921714 0.817502269 0.903265004 0.8181644415  0.900886768
#>    47   10 0.90326500 0.818164442 0.823828793 0.8241718864  0.861095062
#>    18   48 0.79439643 0.711117308 0.813621683 0.6814177339  0.832989095
#>    48   62 0.81362168 0.681417734 0.975855544 0.7212267585  0.908172057
#>    79   87 0.12251144 0.840703868 0.089148972 0.9143497611  0.093112575
#>    87   81 0.08914897 0.914349761 0.063614599 0.8311253732  0.081731145
#>    81   22 0.06361460 0.831125373 0.069479690 0.8156346121  0.078747702
#>    22   28 0.06947969 0.815634612 0.131447986 0.7749064565  0.098280259
#>    28   79 0.13144799 0.774906457 0.122511444 0.8407038682  0.119896595
#>    71   13 0.63246879 0.940458178 0.515307225 0.8776610122  0.592440210
#>    13   65 0.51530722 0.877661012 0.453680392 0.7744574186  0.513519532
#>    65   32 0.45368039 0.774457419 0.502604621 0.7079631577  0.520244053
#>    32   26 0.50260462 0.707963158 0.574245801 0.7057757707  0.539892947
#>    64   91 0.26015811 0.750543371 0.288122852 0.7832834957  0.246141427
#>    75  100 0.28500357 0.870031513 0.221331144 0.8909304833  0.241690667
#>   100   92 0.22133114 0.890930483 0.162302088 0.8669802250  0.202738723
#>    92   79 0.16230209 0.866980225 0.122511444 0.8407038682  0.163788669
#>    47   39 0.90326500 0.818164442 0.858933605 0.9446256620  0.867747653
#>    63   24 0.06504696 0.326183158 0.191697710 0.2554693953  0.124525118
#>    24   33 0.19169771 0.255469395 0.238351715 0.2608308387  0.209714821
#>    33   98 0.23835172 0.260830839 0.266056715 0.3366930966  0.217900888
#>    80   63 0.14018049 0.416342353 0.065046958 0.3261831579  0.114667008
#>    42   41 0.11584153 0.631657109 0.038251256 0.6683095897  0.038232094
#>    41   63 0.03825126 0.668309590 0.065046958 0.3261831579 -0.931188149
#>    25   42 0.11173348 0.471512965 0.115841535 0.6316571093  0.090618446
#>    71   35 0.63246879 0.940458178 0.620414135 1.0000000000  0.549455943
#>    35   82 0.62041414 1.000000000 0.475575969 0.9073757886  0.544858312
#>    82   58 0.47557597 0.907375789 0.442988108 0.9078684974  0.458302884
#>    86   61 0.25129249 0.001279666 0.159926494 0.0228859669  0.175760466
#>    61   57 0.15992649 0.022885967 0.144668076 0.0333743143  0.133224946
#>    57   36 0.14466808 0.033374314 0.101256684 0.1125712440  0.092850631
#>    36   66 0.10125668 0.112571244 0.001881215 0.0001720019  0.050197300
#>    46   53 0.60103925 0.520443712 0.538626588 0.5493443028  0.561245886
#>    53   19 0.53862659 0.549344303 0.454838555 0.3884795079  0.488622498
#>    19   16 0.45483856 0.388479508 0.737046468 0.3439962159  0.590994717
#>    16   34 0.73704647 0.343996216 0.654256981 0.5489428351  0.683628400
#>    34   46 0.65425698 0.548942835 0.601039251 0.5204437122  0.668767256
#>    73   72 0.44625126 0.294866401 0.735408961 0.1664443610  0.602288606
#>    72   99 0.73540896 0.166444361 0.797502659 0.2191768184  0.713275221
#>    99   51 0.79750266 0.219176818 0.851845125 0.2872552070  0.780878030
#>    51   45 0.85184512 0.287255207 0.845916046 0.3318166241  0.787392089
#>    45   90 0.84591605 0.331816624 0.875401223 0.4109329212  0.797844389
#>    90   23 0.87540122 0.410932921 0.787771180 0.4944157961  0.790634065
#>    23   34 0.78777118 0.494415796 0.654256981 0.5489428351  0.689240858
#>    19   73 0.45483856 0.388479508 0.446251258 0.2948664009  0.590053250
#>    63   36 0.06504696 0.326183158 0.101256684 0.1125712440  0.075779051
#>    36   31 0.10125668 0.112571244 0.225828295 0.1548542578  0.153271004
#>    31   33 0.22582830 0.154854258 0.238351715 0.2608308387  0.231979540
#>    26   34 0.57424580 0.705775771 0.654256981 0.5489428351  0.685407752
#>    34   48 0.65425698 0.548942835 0.813621683 0.6814177339  0.714629317
#>    53   78 0.53862659 0.549344303 0.382356276 0.4665283320  0.475239894
#>    21   73 0.33832563 0.350316464 0.446251258 0.2948664009  0.404272949
#>    68   85 0.43701058 0.267384470 0.337648817 0.3198808793  0.360223231
#>    85   33 0.33764882 0.319880879 0.238351715 0.2608308387  0.305019194
#>    31   97 0.22582830 0.154854258 0.366177252 0.0872036163  0.294540047
#>    97   55 0.36617725 0.087203616 0.412245584 0.1306075516  0.347544750
#>    55   68 0.41224558 0.130607552 0.437010577 0.2673844698  0.377621755
#>    98   85 0.26605672 0.336693097 0.337648817 0.3198808793  0.303631784
#>    85   73 0.33764882 0.319880879 0.446251258 0.2948664009  0.398028435
#>    81   49 0.06361460 0.831125373 0.000000000 0.8006657476  0.034105019
#>    49   77 0.00000000 0.800665748 0.029534721 0.7512234236  0.038609859
#>    77   28 0.02953472 0.751223424 0.131447986 0.7749064565  0.080179363
#>    90   76 0.87540122 0.410932921 0.938104380 0.5099860913  0.864490126
#>    76   48 0.93810438 0.509986091 0.813621683 0.6814177339  0.854835157
#>    37   42 0.19103541 0.625035725 0.115841535 0.6316571093  0.146604930
#>    32   43 0.50260462 0.707963158 0.539560375 0.6045277059  0.537053758
#>    43   34 0.53956037 0.604527706 0.654256981 0.5489428351  0.623804203
#>    35   39 0.62041414 1.000000000 0.858933605 0.9446256620  1.194387289
#>    39   96 0.85893360 0.944625662 0.933048282 0.8344800041  0.916242403
#>    96   62 0.93304828 0.834480004 0.975855544 0.7212267585  1.069887110
#>    62   56 0.97585554 0.721226759 1.000000000 0.6036345126  1.667556736
#>    77   93 0.02953472 0.751223424 0.045849438 0.7169532477  0.082291104
#>    93   42 0.04584944 0.716953248 0.115841535 0.6316571093  0.117035440
#>    60   74 0.99777575 0.435553265 0.944908603 0.3691797918  1.154140535
#>    74   89 0.94490860 0.369179792 0.958219025 0.1172698264  1.123363879
#>    89   50 0.95821902 0.117269826 0.909195586 0.0240706549  0.936845731
#>    94   50 0.67687078 0.081091967 0.909195586 0.0240706549  0.733078864
#>    89   83 0.95821902 0.117269826 0.878145092 0.1449405379  0.905783169
#>    83   72 0.87814509 0.144940538 0.735408961 0.1664443610  0.796947100
#>    72   94 0.73540896 0.166444361 0.676870780 0.0810919666  0.781553344
#>    57   59 0.14466808 0.033374314 0.216977153 0.0278631627  0.185641977
#>    59   97 0.21697715 0.027863163 0.366177252 0.0872036163  0.279741261
#>    65   95 0.45368039 0.774457419 0.470848538 0.6257524702  0.414832416
#>    95   43 0.47084854 0.625752470 0.539560375 0.6045277059  0.517507114
#>    43   46 0.53956037 0.604527706 0.601039251 0.5204437122  0.600816074
#>    54   69 0.24100236 0.989597725 0.274594028 0.9733896914  0.421756547
#>    63   66 0.06504696 0.326183158 0.001881215 0.0001720019 -0.098056871
#>    74   51 0.94490860 0.369179792 0.851845125 0.2872552070  0.953132877
#>    99   83 0.79750266 0.219176818 0.878145092 0.1449405379  0.835410475
#>    47   96 0.90326500 0.818164442 0.933048282 0.8344800041  0.886877585
#>    87   54 0.08914897 0.914349761 0.241002364 0.9895977255  0.129749679
#>    49   63 0.00000000 0.800665748 0.065046958 0.3261831579 -1.119031576
#>    95   53 0.47084854 0.625752470 0.538626588 0.5493443028  0.493640973
#>    97   88 0.36617725 0.087203616 0.566053865 0.0326983351  0.431174054
#>    88   67 0.56605386 0.032698335 0.531448691 0.0805114819  0.529841516
#>    67   55 0.53144869 0.080511482 0.412245584 0.1306075516  0.504832521
#>    74   90 0.94490860 0.369179792 0.875401223 0.4109329212  0.891931759
#>    84   96 0.90921714 0.817502269 0.933048282 0.8344800041  0.909037397
#>    76   56 0.93810438 0.509986091 1.000000000 0.6036345126  0.892228775
#>    94   88 0.67687078 0.081091967 0.566053865 0.0326983351  0.731978539
#>    88   86 0.56605386 0.032698335 0.251292492 0.0012796664  0.671207854
#>    78   95 0.38235628 0.466528332 0.470848538 0.6257524702  0.385127124
#>    95   64 0.47084854 0.625752470 0.260158109 0.7505433705  0.352043289
#>    87   70 0.08914897 0.914349761 0.242056269 0.9601491046  0.155258118
#>    70   69 0.24205627 0.960149105 0.274594028 0.9733896914  0.254833629
#>    67   68 0.53144869 0.080511482 0.437010577 0.2673844698  0.504998963
#>    76   60 0.93810438 0.509986091 0.997775748 0.4355532649  1.025901701
#>    61   59 0.15992649 0.022885967 0.216977153 0.0278631627  0.184166602
#>    86   97 0.25129249 0.001279666 0.366177252 0.0872036163  0.283381066
#>    95   91 0.47084854 0.625752470 0.288122852 0.7832834957  0.365789891
#>    72   68 0.73540896 0.166444361 0.437010577 0.2673844698  0.589706635
#>    70   75 0.24205627 0.960149105 0.285003567 0.8700315128  0.276784457
#>    87   92 0.08914897 0.914349761 0.162302088 0.8669802250  0.162531103
#>     1   28 0.23820057 0.730812569 0.131447986 0.7749064565  0.185957638
#>     1   37 0.23820057 0.730812569 0.191035414 0.6250357253  0.163320185
#>     1   52 0.23820057 0.730812569 0.282747345 0.5610843758  0.276749498
#>     1   64 0.23820057 0.730812569 0.260158109 0.7505433705  0.320213137
#>     1    9 0.23820057 0.730812569 0.207649868 0.8092071121  0.222848804
#>     2   11 0.20797469 0.418677087 0.171493140 0.3397613781  0.175164303
#>     2   98 0.20797469 0.418677087 0.266056715 0.3366930966  0.219656161
#>     2   21 0.20797469 0.418677087 0.338325627 0.3503164638  0.288828266
#>     2   78 0.20797469 0.418677087 0.382356276 0.4665283320  0.298070664
#>     2    4 0.20797469 0.418677087 0.208137518 0.4807916769  0.293269811
#>     2   80 0.20797469 0.418677087 0.140180488 0.4163423533  0.172964651
#>     3   71 0.74086069 0.832638247 0.632468787 0.9404581777  0.745640560
#>     3    8 0.74086069 0.832638247 0.581638392 0.7980059757  0.653097200
#>     3   26 0.74086069 0.832638247 0.574245801 0.7057757707  0.676769371
#>     3   18 0.74086069 0.832638247 0.794396431 0.7111173080  0.683679414
#>     3   10 0.74086069 0.832638247 0.823828793 0.8241718864  0.776997685
#>     3   39 0.74086069 0.832638247 0.858933605 0.9446256620  0.789598523
#>     4   52 0.20813752 0.480791677 0.282747345 0.5610843758  0.295322870
#>     4   37 0.20813752 0.480791677 0.191035414 0.6250357253  0.209736190
#>     4   25 0.20813752 0.480791677 0.111733484 0.4715129652  0.153078091
#>     5   65 0.33058535 0.857917828 0.453680392 0.7744574186  0.371248982
#>     5   58 0.33058535 0.857917828 0.442988108 0.9078684974  0.406810132
#>     5   69 0.33058535 0.857917828 0.274594028 0.9733896914  0.359882693
#>     5   75 0.33058535 0.857917828 0.285003567 0.8700315128  0.324330086
#>     5   91 0.33058535 0.857917828 0.288122852 0.7832834957  0.297986315
#>     6   84 0.86463926 0.737453470 0.909217141 0.8175022692  0.924239264
#>     6   47 0.86463926 0.737453470 0.903265004 0.8181644415  0.900886768
#>     6   10 0.86463926 0.737453470 0.823828793 0.8241718864  0.861095062
#>     6   18 0.86463926 0.737453470 0.794396431 0.7111173080  0.813709900
#>     6   48 0.86463926 0.737453470 0.813621683 0.6814177339  0.832989095
#>     6   62 0.86463926 0.737453470 0.975855544 0.7212267585  0.908172057
#>     7   87 0.09416720 0.829010331 0.089148972 0.9143497611  0.093112575
#>     7   81 0.09416720 0.829010331 0.063614599 0.8311253732  0.081731145
#>     7   22 0.09416720 0.829010331 0.069479690 0.8156346121  0.078747702
#>     7   28 0.09416720 0.829010331 0.131447986 0.7749064565  0.098280259
#>     7   79 0.09416720 0.829010331 0.122511444 0.8407038682  0.119896595
#>     8   13 0.58163839 0.798005976 0.515307225 0.8776610122  0.592440210
#>     8   65 0.58163839 0.798005976 0.453680392 0.7744574186  0.513519532
#>     8   32 0.58163839 0.798005976 0.502604621 0.7079631577  0.520244053
#>     9   91 0.20764987 0.809207112 0.288122852 0.7832834957  0.246141427
#>     9   75 0.20764987 0.809207112 0.285003567 0.8700315128  0.257344914
#>     9  100 0.20764987 0.809207112 0.221331144 0.8909304833  0.241690667
#>     9   92 0.20764987 0.809207112 0.162302088 0.8669802250  0.202738723
#>     9   79 0.20764987 0.809207112 0.122511444 0.8407038682  0.163788669
#>    11   24 0.17149314 0.339761378 0.191697710 0.2554693953  0.124525118
#>    11   33 0.17149314 0.339761378 0.238351715 0.2608308387  0.209714821
#>    11   63 0.17149314 0.339761378 0.065046958 0.3261831579  0.114667008
#>    12   41 0.06814813 0.471879939 0.038251256 0.6683095897  0.038232094
#>    12   63 0.06814813 0.471879939 0.065046958 0.3261831579 -0.931188149
#>    12   80 0.06814813 0.471879939 0.140180488 0.4163423533  0.069362169
#>    12   25 0.06814813 0.471879939 0.111733484 0.4715129652  0.089548943
#>    12   42 0.06814813 0.471879939 0.115841535 0.6316571093  0.090618446
#>    13   35 0.51530722 0.877661012 0.620414135 1.0000000000  0.549455943
#>    13   82 0.51530722 0.877661012 0.475575969 0.9073757886  0.544858312
#>    13   58 0.51530722 0.877661012 0.442988108 0.9078684974  0.458302884
#>    14   61 0.09830826 0.000000000 0.159926494 0.0228859669  0.175760466
#>    14   57 0.09830826 0.000000000 0.144668076 0.0333743143  0.133224946
#>    14   36 0.09830826 0.000000000 0.101256684 0.1125712440  0.092850631
#>    14   66 0.09830826 0.000000000 0.001881215 0.0001720019  0.050197300
#>    15   53 0.57945432 0.480730595 0.538626588 0.5493443028  0.561245886
#>    15   19 0.57945432 0.480730595 0.454838555 0.3884795079  0.488622498
#>    15   16 0.57945432 0.480730595 0.737046468 0.3439962159  0.590994717
#>    15   34 0.57945432 0.480730595 0.654256981 0.5489428351  0.683628400
#>    15   46 0.57945432 0.480730595 0.601039251 0.5204437122  0.668767256
#>    16   72 0.73704647 0.343996216 0.735408961 0.1664443610  0.602288606
#>    16   99 0.73704647 0.343996216 0.797502659 0.2191768184  0.713275221
#>    16   51 0.73704647 0.343996216 0.851845125 0.2872552070  0.780878030
#>    16   45 0.73704647 0.343996216 0.845916046 0.3318166241  0.787392089
#>    16   90 0.73704647 0.343996216 0.875401223 0.4109329212  0.797844389
#>    16   23 0.73704647 0.343996216 0.787771180 0.4944157961  0.790634065
#>    16   73 0.73704647 0.343996216 0.446251258 0.2948664009  0.590053250
#>    17   36 0.18366055 0.230488665 0.101256684 0.1125712440  0.075779051
#>    17   31 0.18366055 0.230488665 0.225828295 0.1548542578  0.153271004
#>    17   33 0.18366055 0.230488665 0.238351715 0.2608308387  0.231979540
#>    17   24 0.18366055 0.230488665 0.191697710 0.2554693953  0.217884978
#>    17   63 0.18366055 0.230488665 0.065046958 0.3261831579  0.114756922
#>    18   34 0.79439643 0.711117308 0.654256981 0.5489428351  0.685407752
#>    19   78 0.45483856 0.388479508 0.382356276 0.4665283320  0.475239894
#>    19   21 0.45483856 0.388479508 0.338325627 0.3503164638  0.387123647
#>    20   85 0.29408033 0.195967076 0.337648817 0.3198808793  0.360223231
#>    20   33 0.29408033 0.195967076 0.238351715 0.2608308387  0.305019194
#>    20   31 0.29408033 0.195967076 0.225828295 0.1548542578  0.241056671
#>    20   97 0.29408033 0.195967076 0.366177252 0.0872036163  0.294540047
#>    20   55 0.29408033 0.195967076 0.412245584 0.1306075516  0.347544750
#>    20   68 0.29408033 0.195967076 0.437010577 0.2673844698  0.377621755
#>    21   85 0.33832563 0.350316464 0.337648817 0.3198808793  0.303631784
#>    22   49 0.06947969 0.815634612 0.000000000 0.8006657476  0.034105019
#>    22   77 0.06947969 0.815634612 0.029534721 0.7512234236  0.038609859
#>    23   76 0.78777118 0.494415796 0.938104380 0.5099860913  0.864490126
#>    23   48 0.78777118 0.494415796 0.813621683 0.6814177339  0.854835157
#>    26   43 0.57424580 0.705775771 0.539560375 0.6045277059  0.537053758
#>    27   39 0.98259466 0.927663307 0.858933605 0.9446256620  1.194387289
#>    27   96 0.98259466 0.927663307 0.933048282 0.8344800041  0.916242403
#>    27   62 0.98259466 0.927663307 0.975855544 0.7212267585  1.069887110
#>    27   56 0.98259466 0.927663307 1.000000000 0.6036345126  1.667556736
#>    28   93 0.13144799 0.774906457 0.045849438 0.7169532477  0.082291104
#>    28   42 0.13144799 0.774906457 0.115841535 0.6316571093  0.117035440
#>    29   74 0.98546270 0.089545264 0.944908603 0.3691797918  1.154140535
#>    29   89 0.98546270 0.089545264 0.958219025 0.1172698264  1.123363879
#>    29   50 0.98546270 0.089545264 0.909195586 0.0240706549  0.936845731
#>    30   50 0.88261083 0.043277687 0.909195586 0.0240706549  0.733078864
#>    30   89 0.88261083 0.043277687 0.958219025 0.1172698264  0.925682071
#>    30   83 0.88261083 0.043277687 0.878145092 0.1449405379  0.905783169
#>    30   72 0.88261083 0.043277687 0.735408961 0.1664443610  0.796947100
#>    30   94 0.88261083 0.043277687 0.676870780 0.0810919666  0.781553344
#>    31   57 0.22582830 0.154854258 0.144668076 0.0333743143  0.174559396
#>    31   59 0.22582830 0.154854258 0.216977153 0.0278631627  0.185641977
#>    32   95 0.50260462 0.707963158 0.470848538 0.6257524702  0.414832416
#>    35   69 0.62041414 1.000000000 0.274594028 0.9733896914  0.421756547
#>    35   58 0.62041414 1.000000000 0.442988108 0.9078684974  0.435816082
#>    38   74 0.89014939 0.183898554 0.944908603 0.3691797918  1.018316204
#>    38   51 0.89014939 0.183898554 0.851845125 0.2872552070  0.953132877
#>    38   99 0.89014939 0.183898554 0.797502659 0.2191768184  0.854452823
#>    38   83 0.89014939 0.183898554 0.878145092 0.1449405379  0.835410475
#>    38   89 0.89014939 0.183898554 0.958219025 0.1172698264  0.925311174
#>    40   81 0.07777007 0.919298432 0.063614599 0.8311253732 -0.002197272
#>    40   87 0.07777007 0.919298432 0.089148972 0.9143497611  0.065710270
#>    40   54 0.07777007 0.919298432 0.241002364 0.9895977255  0.129749679
#>    41   77 0.03825126 0.668309590 0.029534721 0.7512234236 -0.026800359
#>    41   49 0.03825126 0.668309590 0.000000000 0.8006657476 -0.123762978
#>    41   93 0.03825126 0.668309590 0.045849438 0.7169532477  0.093403509
#>    43   53 0.53956037 0.604527706 0.538626588 0.5493443028  0.493640973
#>    44   97 0.51541452 0.077670122 0.366177252 0.0872036163  0.439103891
#>    44   88 0.51541452 0.077670122 0.566053865 0.0326983351  0.431174054
#>    44   67 0.51541452 0.077670122 0.531448691 0.0805114819  0.529841516
#>    44   55 0.51541452 0.077670122 0.412245584 0.1306075516  0.504832521
#>    45   74 0.84591605 0.331816624 0.944908603 0.3691797918  0.907908557
#>    48   56 0.81362168 0.681417734 1.000000000 0.6036345126  0.892228775
#>    50   88 0.90919559 0.024070655 0.566053865 0.0326983351  0.731978539
#>    50   86 0.90919559 0.024070655 0.251292492 0.0012796664  0.671207854
#>    52   95 0.28274735 0.561084376 0.470848538 0.6257524702  0.385127124
#>    54   70 0.24100236 0.989597725 0.242056269 0.9601491046  0.155258118
#>    56   60 1.00000000 0.603634513 0.997775748 0.4355532649  1.025901701
#>    59   86 0.21697715 0.027863163 0.251292492 0.0012796664  0.193922824
#>    60   90 0.99777575 0.435553265 0.875401223 0.4109329212  0.932362565
#>    67   94 0.53144869 0.080511482 0.676870780 0.0810919666  0.604096428
#>    67   72 0.53144869 0.080511482 0.735408961 0.1664443610  0.603707798
#>    68   73 0.43701058 0.267384470 0.446251258 0.2948664009  0.590946991
#>    70   92 0.24205627 0.960149105 0.162302088 0.8669802250  0.162531103
#>    70  100 0.24205627 0.960149105 0.221331144 0.8909304833  0.164787581
#>    14   86 0.09830826 0.000000000 0.251292492 0.0012796664  0.175760466
#>    27   35 0.98259466 0.927663307 0.620414135 1.0000000000  1.194387289
#>    29   60 0.98546270 0.089545264 0.997775748 0.4355532649  1.154140535
#>    35   54 0.62041414 1.000000000 0.241002364 0.9895977255  0.421756547
#>    40   49 0.07777007 0.919298432 0.000000000 0.8006657476 -0.002197272
#>    49   66 0.00000000 0.800665748 0.001881215 0.0001720019 -1.176702312
#>            my1          mx2           my2 bp1 bp2
#>   0.7556034102  0.160415969  8.123464e-01   0   0
#>   0.7007974803  0.159691281  6.993547e-01   0   0
#>   0.6502201424  0.209736190  5.541171e-01   0   0
#>   0.6616275872  0.352043289  6.654227e-01   0   0
#>   0.7699800613  0.246141427  7.908286e-01   0   0
#>   0.3859545459  0.114667008  3.612182e-01   0   0
#>   0.3653866258  0.217900888  3.112896e-01   0   0
#>   0.4143919973  0.303631784  3.358626e-01   0   0
#>   0.4320155233  0.387123647  3.982749e-01   0   0
#>   0.4495110026  0.295322870  4.745880e-01   0   0
#>   0.4498263706  0.161283885  4.621429e-01   0   0
#>   0.9458368947  0.744750903  9.941816e-01   0   0
#>   0.8528026017  0.592440210  8.744465e-01   0   0
#>   0.7439695027  0.539892947  7.549407e-01   0   0
#>   0.7348941945  0.685407752  6.636610e-01   0   0
#>   0.7760053214  0.813709900  7.664478e-01   0   0
#>   0.8994902153  0.867747653  8.767146e-01   0   0
#>   0.4745879725  0.385127124  5.691912e-01   0   0
#>   0.5473994974  0.146604930  5.507432e-01   0   0
#>   0.4621428719  0.089548943  4.251550e-01   0   0
#>   0.7853861551  0.366110724  6.890039e-01   0   0
#>   0.8378349910  0.457914457  8.419308e-01   0   0
#>   0.9434346491  0.435816082  1.138589e+00   0   0
#>   0.9261954933  0.276784457  9.214070e-01   0   0
#>   0.8270682563  0.257344914  8.256069e-01   0   0
#>   0.7566999420  0.955220068  7.781437e-01   0   0
#>   0.7697045687  0.909037397  8.429690e-01   0   0
#>   0.7887476335  0.867747653  8.767146e-01   0   0
#>   0.7150270004  0.714629317  6.384099e-01   0   0
#>   0.6465768879  0.908293996  6.460799e-01   0   0
#>   0.8717655744  0.121945289  8.848271e-01   0   0
#>   0.8710963110  0.065710270  8.760117e-01   0   0
#>   0.8279993517  0.034105019  8.110968e-01   0   0
#>   0.7919481969  0.080179363  7.644075e-01   0   0
#>   0.8068431397  0.160415969  8.123464e-01   0   0
#>   0.8744464843  0.549455943  9.546428e-01   0   0
#>   0.8087268397  0.457914457  8.419308e-01   0   0
#>   0.7721871829  0.414832416  6.946289e-01   0   0
#>   0.7549406529  0.537053758  6.619517e-01   0   0
#>   0.7908286237  0.365789891  6.886317e-01   0   0
#>   0.8455152299  0.264711940  9.156536e-01   0   0
#>   0.8520361588  0.164787581  9.455725e-01   0   0
#>   0.8214631400  0.121945289  8.848271e-01   0   0
#>   0.8767145864  0.886877585  8.834206e-01   0   0
#>   0.2839357814  0.114756922  2.664406e-01   0   0
#>   0.3043555333  0.217884978  2.332608e-01   0   0
#>   0.3112895893  0.291545351  2.843945e-01   0   0
#>   0.3612182357  0.069362169  3.989727e-01   0   0
#>   0.5678166746  0.093403509  6.846100e-01   0   0
#>   0.4202695264 -1.119031576  4.055574e-01   0   0
#>   0.5521793748  0.146604930  5.507432e-01   0   0
#>   0.9546428411  0.744750903  9.941816e-01   0   0
#>   0.9585928720  0.462011391  1.088142e+00   0   0
#>   0.8428606927  0.462011391  1.088142e+00   0   0
#>  -0.1141389843  0.193922824 -3.733629e-02   0   0
#>   0.0003837559  0.184166602  7.449353e-02   0   0
#>   0.0564671781  0.174559396  1.012554e-01   0   0
#>   0.0575843380 -0.098056871  1.886602e-01   0   0
#>   0.5163497631  0.588910395  5.760930e-01   0   0
#>   0.4731361179  0.475239894  4.801066e-01   0   0
#>   0.3348483973  0.590053250  3.288756e-01   0   0
#>   0.4416126274  0.689240858  4.438798e-01   0   0
#>   0.4579096007  0.600816074  5.847980e-01   0   0
#>   0.2564555677  0.590946991  2.309186e-01   0   0
#>   0.2554319723  0.805451278  1.468925e-01   0   0
#>   0.2881753470  0.854452823  2.294455e-01   0   0
#>   0.3013546202  0.907908557  3.173898e-01   0   0
#>   0.3947844772  0.891931759  3.597199e-01   0   0
#>   0.4096878484  0.864490126  4.872128e-01   0   0
#>   0.4438798155  0.750877912  5.948035e-01   0   0
#>   0.3288756048  0.404272949  3.459176e-01   0   0
#>   0.2181274299 -0.098056871  1.886602e-01   0   0
#>   0.1639739675  0.174559396  1.012554e-01   0   0
#>   0.2078556021  0.241056671  2.067829e-01   0   0
#>   0.6636610170  0.623804203  6.322328e-01   0   0
#>   0.6384098537  0.750877912  5.948035e-01   0   0
#>   0.4801065812  0.445940866  5.353926e-01   0   0
#>   0.3459175636  0.398028435  3.337635e-01   0   0
#>   0.2423273343  0.389909941  2.985164e-01   0   0
#>   0.2617372480  0.291545351  2.843945e-01   0   0
#>   0.1179943449  0.279741261  8.729258e-02   0   0
#>   0.1531300034  0.439103891  5.595041e-02   0   0
#>   0.2075070327  0.504998963  1.844440e-01   0   0
#>   0.3358626492  0.291545351  2.843945e-01   0   0
#>   0.3337635076  0.389909941  2.985164e-01   0   0
#>   0.8110967983 -0.002197272  8.869137e-01   0   0
#>   0.7901870702 -0.123762978  6.931925e-01   0   0
#>   0.7644075030  0.082291104  7.553202e-01   0   0
#>   0.4872128468  0.932362565  4.442479e-01   0   0
#>   0.5804328144  0.892228775  6.075857e-01   0   0
#>   0.5507431980  0.159691281  6.993547e-01   0   0
#>   0.6619516952  0.517507114  6.549680e-01   0   0
#>   0.6322328439  0.600816074  5.847980e-01   0   0
#>   2.9309457910  0.744750903  9.941816e-01   0   0
#>   0.9031796109  0.886877585  8.834206e-01   0   0
#>   0.8214853729  0.955220068  7.781437e-01   0   0
#>   0.8019744585  0.908293996  6.460799e-01   0   0
#>   0.7553202208 -0.026800359  7.033860e-01   0   0
#>   0.7040018423  0.093403509  6.846100e-01   0   0
#>   0.2567657787  0.934958112  4.313468e-01   0   0
#>   0.2523023829  1.018316204  2.467519e-01   0   0
#>   0.0690194061  0.925682071  7.489157e-02   0   0
#>  -0.1916935938  0.731978539 -1.961767e-01   0   0
#>   0.0952250850  0.925311174  1.517355e-01   0   0
#>   0.0904442523  0.805451278  1.468925e-01   0   0
#>   0.0720465031  0.603707798  1.940203e-01   0   0
#>   0.0938511957  0.184166602  7.449353e-02   0   0
#>   0.0872925772  0.283381066  7.814099e-02   0   0
#>   0.6946288633  0.366110724  6.890039e-01   0   0
#>   0.6549680069  0.493640973  5.777051e-01   0   0
#>   0.5847979869  0.588910395  5.760930e-01   0   0
#>   1.3213025811  0.254833629  9.753495e-01   0   0
#>   0.1886602016 -1.176702312  3.976513e-01   0   0
#>   0.2660165728  0.907908557  3.173898e-01   0   0
#>   0.1794370153  0.805451278  1.468925e-01   0   0
#>   0.8834206471  0.909037397  8.429690e-01   0   0
#>   1.0232629965  0.155258118  9.717860e-01   0   0
#>   0.4055574316 -1.176702312  3.976513e-01   0   0
#>   0.5777051292  0.445940866  5.353926e-01   0   0
#>  -0.0681832031  0.420926041 -1.057637e-01   0   0
#>   0.0429188071  0.604096428  9.666143e-02   0   0
#>   0.1840479150  0.504998963  1.844440e-01   0   0
#>   0.3597198591  0.934958112  4.313468e-01   0   0
#>   0.8429689702  0.955220068  7.781437e-01   0   0
#>   0.6075856685  1.025901701  5.192364e-01   0   0
#>  -0.1961767067  0.604096428  9.666143e-02   0   0
#>  -2.6131598257  0.420926041 -1.057637e-01   0   0
#>   0.5691911875  0.445940866  5.353926e-01   0   0
#>   0.6654227031  0.365789891  6.886317e-01   0   0
#>   0.9717859501  0.162531103  9.475041e-01   0   0
#>   0.9753495484  0.276784457  9.214070e-01   0   0
#>   0.1844439618  0.589706635  2.272518e-01   0   0
#>   0.5192364096  0.932362565  4.442479e-01   0   0
#>   0.0744935314  0.193922824 -3.733629e-02   0   0
#>   0.0781409900  0.420926041 -1.057637e-01   0   0
#>   0.6886317460  0.366110724  6.890039e-01   0   0
#>   0.2272518218  0.590946991  2.309186e-01   0   0
#>   0.9214070180  0.264711940  9.156536e-01   0   0
#>   0.9475041060  0.121945289  8.848271e-01   0   0
#>   0.7556034102  0.163320185  7.007975e-01   0   0
#>   0.7007974803  0.276749498  6.502201e-01   0   0
#>   0.6502201424  0.320213137  6.616276e-01   0   0
#>   0.6616275872  0.222848804  7.699801e-01   0   0
#>   0.7699800613  0.185957638  7.556034e-01   0   0
#>   0.3859545459  0.219656161  3.653866e-01   0   0
#>   0.3653866258  0.288828266  4.143920e-01   0   0
#>   0.4143919973  0.298070664  4.320155e-01   0   0
#>   0.4320155233  0.293269811  4.495110e-01   0   0
#>   0.4495110026  0.172964651  4.498264e-01   0   0
#>   0.4498263706  0.175164303  3.859545e-01   0   0
#>   0.9458368947  0.653097200  8.528026e-01   0   0
#>   0.8528026017  0.676769371  7.439695e-01   0   0
#>   0.7439695027  0.683679414  7.348942e-01   0   0
#>   0.7348941945  0.776997685  7.760053e-01   0   0
#>   0.7760053214  0.789598523  8.994902e-01   0   0
#>   0.8994902153  0.745640560  9.458369e-01   0   0
#>   0.4745879725  0.209736190  5.541171e-01   0   0
#>   0.5541170896  0.153078091  5.473995e-01   0   0
#>   0.5473994974  0.161283885  4.621429e-01   0   0
#>   0.7853861551  0.406810132  8.378350e-01   0   0
#>   0.8378349910  0.359882693  9.434346e-01   0   0
#>   0.9434346491  0.324330086  9.261955e-01   0   0
#>   0.9261954933  0.297986315  8.270683e-01   0   0
#>   0.8270682563  0.371248982  7.853862e-01   0   0
#>   0.7566999420  0.900886768  7.697046e-01   0   0
#>   0.7697045687  0.861095062  7.887476e-01   0   0
#>   0.7887476335  0.813709900  7.664478e-01   0   0
#>   0.7664477500  0.832989095  7.150270e-01   0   0
#>   0.7150270004  0.908172057  6.465769e-01   0   0
#>   0.6465768879  0.924239264  7.566999e-01   0   0
#>   0.8717655744  0.081731145  8.710963e-01   0   0
#>   0.8710963110  0.078747702  8.279994e-01   0   0
#>   0.8279993517  0.098280259  7.919482e-01   0   0
#>   0.7919481969  0.119896595  8.068431e-01   0   0
#>   0.8068431397  0.093112575  8.717656e-01   0   0
#>   0.8744464843  0.513519532  8.087268e-01   0   0
#>   0.8087268397  0.520244053  7.721872e-01   0   0
#>   0.7721871829  0.539892947  7.549407e-01   0   0
#>   0.7908286237  0.257344914  8.256069e-01   0   0
#>   0.8256068728  0.241690667  8.455152e-01   0   0
#>   0.8455152299  0.202738723  8.520362e-01   0   0
#>   0.8520361588  0.163788669  8.214631e-01   0   0
#>   0.8214631400  0.160415969  8.123464e-01   0   0
#>   0.2839357814  0.209714821  3.043555e-01   0   0
#>   0.3043555333  0.217900888  3.112896e-01   0   0
#>   0.3612182357  0.124525118  2.839358e-01   0   0
#>   0.5678166746 -0.931188149  4.202695e-01   0   0
#>   0.4202695264  0.069362169  3.989727e-01   0   0
#>   0.3989727033  0.089548943  4.251550e-01   0   0
#>   0.4251549885  0.090618446  5.521794e-01   0   0
#>   0.5521793748  0.038232094  5.678167e-01   0   0
#>   0.9546428411  0.544858312  9.585929e-01   0   0
#>   0.9585928720  0.458302884  8.428607e-01   0   0
#>   0.8428606927  0.457914457  8.419308e-01   0   0
#>  -0.1141389843  0.133224946  3.837559e-04   0   0
#>   0.0003837559  0.092850631  5.646718e-02   0   0
#>   0.0564671781  0.050197300  5.758434e-02   0   0
#>   0.0575843380  0.031698903 -1.031292e+01   0   1
#>   0.5163497631  0.488622498  4.731361e-01   0   0
#>   0.4731361179  0.590994717  3.348484e-01   0   0
#>   0.3348483973  0.683628400  4.416126e-01   0   0
#>   0.4416126274  0.668767256  4.579096e-01   0   0
#>   0.4579096007  0.561245886  5.163498e-01   0   0
#>   0.2564555677  0.713275221  2.554320e-01   0   0
#>   0.2554319723  0.780878030  2.881753e-01   0   0
#>   0.2881753470  0.787392089  3.013546e-01   0   0
#>   0.3013546202  0.797844389  3.947845e-01   0   0
#>   0.3947844772  0.790634065  4.096878e-01   0   0
#>   0.4096878484  0.689240858  4.438798e-01   0   0
#>   0.3288756048  0.602288606  2.564556e-01   0   0
#>   0.2181274299  0.153271004  1.639740e-01   0   0
#>   0.1639739675  0.231979540  2.078556e-01   0   0
#>   0.2078556021  0.217884978  2.332608e-01   0   0
#>   0.2332607706  0.114756922  2.664406e-01   0   0
#>   0.2664406118  0.075779051  2.181274e-01   0   0
#>   0.6636610170  0.714629317  6.384099e-01   0   0
#>   0.4801065812  0.387123647  3.982749e-01   0   0
#>   0.3982749020  0.404272949  3.459176e-01   0   0
#>   0.2423273343  0.305019194  2.617372e-01   0   0
#>   0.2617372480  0.241056671  2.067829e-01   0   0
#>   0.2067829431  0.294540047  1.179943e-01   0   0
#>   0.1179943449  0.347544750  1.531300e-01   0   0
#>   0.1531300034  0.377621755  2.075070e-01   0   0
#>   0.2075070327  0.360223231  2.423273e-01   0   0
#>   0.3358626492  0.398028435  3.337635e-01   0   0
#>   0.8110967983  0.038609859  7.901871e-01   0   0
#>   0.7901870702  0.080179363  7.644075e-01   0   0
#>   0.4872128468  0.854835157  5.804328e-01   0   0
#>   0.5804328144  0.750877912  5.948035e-01   0   0
#>   0.6619516952  0.623804203  6.322328e-01   0   0
#>   2.9309457910  0.916242403  9.031796e-01   0   0
#>   0.9031796109  1.069887110  8.214854e-01   0   0
#>   0.8214853729  1.667556736  8.019745e-01   0   0
#>   0.8019744585  4.744823247  9.672711e-01   0   1
#>   0.7553202208  0.117035440  7.040018e-01   0   0
#>   0.7040018423  0.159691281  6.993547e-01   0   0
#>   0.2567657787  1.123363879  2.523024e-01   0   0
#>   0.2523023829  0.936845731  6.901941e-02   0   0
#>   0.0690194061  7.417172130 -7.479492e+00   0   1
#>  -0.1916935938  0.925682071  7.489157e-02   0   0
#>   0.0748915719  0.905783169  9.522509e-02   0   0
#>   0.0952250850  0.796947100  9.044425e-02   0   0
#>   0.0904442523  0.781553344  7.204650e-02   0   0
#>   0.0720465031  0.733078864 -1.916936e-01   0   0
#>   0.1012554199  0.185641977  9.385120e-02   0   0
#>   0.0938511957  0.279741261  8.729258e-02   0   0
#>   0.6946288633  0.517507114  6.549680e-01   0   0
#>   1.3213025811  0.435816082  1.138589e+00   0   0
#>   1.1385888029  0.462011391  1.088142e+00   0   0
#>   0.2467518726  0.953132877  2.660166e-01   0   0
#>   0.2660165728  0.854452823  2.294455e-01   0   0
#>   0.2294454656  0.835410475  1.794370e-01   0   0
#>   0.1794370153  0.925311174  1.517355e-01   0   0
#>   0.1517355266  1.018316204  2.467519e-01   0   0
#>   0.8869137402  0.065710270  8.760117e-01   0   0
#>   0.8760117330  0.129749679  1.023263e+00   0   0
#>   1.0232629965 -2.095843134  6.191005e+00   0   1
#>   0.7033859592 -0.123762978  6.931925e-01   0   0
#>   0.6931925093 -1.119031576  4.055574e-01   0   0
#>   0.6846100099 -0.026800359  7.033860e-01   0   0
#>   0.5777051292  0.588910395  5.760930e-01   0   0
#>   0.0559504095  0.431174054 -6.818320e-02   0   0
#>  -0.0681832031  0.529841516  4.291881e-02   0   0
#>   0.0429188071  0.504832521  1.840479e-01   0   0
#>   0.1840479150  0.439103891  5.595041e-02   0   0
#>   0.3173898280  0.891931759  3.597199e-01   0   0
#>   0.6075856685  0.908293996  6.460799e-01   0   0
#>  -0.1961767067  0.671207854 -2.613160e+00   0   0
#>  -2.6131598257  0.723799703 -4.131319e+00   0   1
#>   0.5691911875  0.352043289  6.654227e-01   0   0
#>   0.9717859501  0.254833629  9.753495e-01   0   0
#>   0.5192364096  6.974363693  4.405193e-01   0   1
#>  -0.0373362938  0.283381066  7.814099e-02   0   0
#>   0.4442478796  0.934958112  4.313468e-01   0   0
#>   0.0966614340  0.603707798  1.940203e-01   0   0
#>   0.1940202907  0.589706635  2.272518e-01   0   0
#>   0.2309185528  0.389909941  2.985164e-01   0   0
#>   0.9475041060  0.164787581  9.455725e-01   0   0
#>   0.9455725217  0.264711940  9.156536e-01   0   0
#>  -0.1141389843  0.230433482 -6.650303e+00   0   1
#>   2.9309457910  1.724686452  5.586086e+00   0   1
#>   0.2567657787  4.040591817  1.540485e-01   0   1
#>   1.3213025811  0.349549368  3.954982e+00   0   1
#>   0.8869137402 -5.897897861  4.751861e+00   0   1
#>   0.3976513335 -2.425924410  3.947156e-01   0   1
#> 
#> $x
#>               [,1]         [,2]
#>   [1,] 0.238200569 0.7308125686
#>   [2,] 0.207974691 0.4186770868
#>   [3,] 0.740860690 0.8326382472
#>   [4,] 0.208137518 0.4807916769
#>   [5,] 0.330585349 0.8579178283
#>   [6,] 0.864639259 0.7374534701
#>   [7,] 0.094167195 0.8290103306
#>   [8,] 0.581638392 0.7980059757
#>   [9,] 0.207649868 0.8092071121
#>  [10,] 0.823828793 0.8241718864
#>  [11,] 0.171493140 0.3397613781
#>  [12,] 0.068148130 0.4718799394
#>  [13,] 0.515307225 0.8776610122
#>  [14,] 0.098308260 0.0000000000
#>  [15,] 0.579454323 0.4807305955
#>  [16,] 0.737046468 0.3439962159
#>  [17,] 0.183660551 0.2304886653
#>  [18,] 0.794396431 0.7111173080
#>  [19,] 0.454838555 0.3884795079
#>  [20,] 0.294080334 0.1959670757
#>  [21,] 0.338325627 0.3503164638
#>  [22,] 0.069479690 0.8156346121
#>  [23,] 0.787771180 0.4944157961
#>  [24,] 0.191697710 0.2554693953
#>  [25,] 0.111733484 0.4715129652
#>  [26,] 0.574245801 0.7057757707
#>  [27,] 0.982594662 0.9276633066
#>  [28,] 0.131447986 0.7749064565
#>  [29,] 0.985462702 0.0895452642
#>  [30,] 0.882610827 0.0432776872
#>  [31,] 0.225828295 0.1548542578
#>  [32,] 0.502604621 0.7079631577
#>  [33,] 0.238351715 0.2608308387
#>  [34,] 0.654256981 0.5489428351
#>  [35,] 0.620414135 1.0000000000
#>  [36,] 0.101256684 0.1125712440
#>  [37,] 0.191035414 0.6250357253
#>  [38,] 0.890149392 0.1838985535
#>  [39,] 0.858933605 0.9446256620
#>  [40,] 0.077770071 0.9192984322
#>  [41,] 0.038251256 0.6683095897
#>  [42,] 0.115841535 0.6316571093
#>  [43,] 0.539560375 0.6045277059
#>  [44,] 0.515414517 0.0776701220
#>  [45,] 0.845916046 0.3318166241
#>  [46,] 0.601039251 0.5204437122
#>  [47,] 0.903265004 0.8181644415
#>  [48,] 0.813621683 0.6814177339
#>  [49,] 0.000000000 0.8006657476
#>  [50,] 0.909195586 0.0240706549
#>  [51,] 0.851845125 0.2872552070
#>  [52,] 0.282747345 0.5610843758
#>  [53,] 0.538626588 0.5493443028
#>  [54,] 0.241002364 0.9895977255
#>  [55,] 0.412245584 0.1306075516
#>  [56,] 1.000000000 0.6036345126
#>  [57,] 0.144668076 0.0333743143
#>  [58,] 0.442988108 0.9078684974
#>  [59,] 0.216977153 0.0278631627
#>  [60,] 0.997775748 0.4355532649
#>  [61,] 0.159926494 0.0228859669
#>  [62,] 0.975855544 0.7212267585
#>  [63,] 0.065046958 0.3261831579
#>  [64,] 0.260158109 0.7505433705
#>  [65,] 0.453680392 0.7744574186
#>  [66,] 0.001881215 0.0001720019
#>  [67,] 0.531448691 0.0805114819
#>  [68,] 0.437010577 0.2673844698
#>  [69,] 0.274594028 0.9733896914
#>  [70,] 0.242056269 0.9601491046
#>  [71,] 0.632468787 0.9404581777
#>  [72,] 0.735408961 0.1664443610
#>  [73,] 0.446251258 0.2948664009
#>  [74,] 0.944908603 0.3691797918
#>  [75,] 0.285003567 0.8700315128
#>  [76,] 0.938104380 0.5099860913
#>  [77,] 0.029534721 0.7512234236
#>  [78,] 0.382356276 0.4665283320
#>  [79,] 0.122511444 0.8407038682
#>  [80,] 0.140180488 0.4163423533
#>  [81,] 0.063614599 0.8311253732
#>  [82,] 0.475575969 0.9073757886
#>  [83,] 0.878145092 0.1449405379
#>  [84,] 0.909217141 0.8175022692
#>  [85,] 0.337648817 0.3198808793
#>  [86,] 0.251292492 0.0012796664
#>  [87,] 0.089148972 0.9143497611
#>  [88,] 0.566053865 0.0326983351
#>  [89,] 0.958219025 0.1172698264
#>  [90,] 0.875401223 0.4109329212
#>  [91,] 0.288122852 0.7832834957
#>  [92,] 0.162302088 0.8669802250
#>  [93,] 0.045849438 0.7169532477
#>  [94,] 0.676870780 0.0810919666
#>  [95,] 0.470848538 0.6257524702
#>  [96,] 0.933048282 0.8344800041
#>  [97,] 0.366177252 0.0872036163
#>  [98,] 0.266056715 0.3366930966
#>  [99,] 0.797502659 0.2191768184
#> [100,] 0.221331144 0.8909304833
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.2382006,0.7308126) [5]: 9 28 37 52 64 
#> 2: (0.2079747,0.4186771) [6]: 4 11 21 78 80 98 
#> 3: (0.7408607,0.8326382) [6]: 8 10 18 26 39 71 
#> 4: (0.2081375,0.4807917) [6]: 2 25 37 52 78 80 
#> 5: (0.3305853,0.8579178) [5]: 58 65 69 75 91 
#> 6: (0.8646393,0.7374535) [6]: 10 18 47 48 62 84 
#> 7: (0.0941672,0.8290103) [5]: 22 28 79 81 87 
#> 8: (0.5816384,0.798006) [6]: 3 13 26 32 65 71 
#> 9: (0.2076499,0.8092071) [8]: 1 28 64 75 79 91 92 100 
#> 10: (0.8238288,0.8241719) [5]: 3 6 18 39 47 
#> 11: (0.1714931,0.3397614) [6]: 2 24 33 63 80 98 
#> 12: (0.06814813,0.4718799) [5]: 25 41 42 63 80 
#> 13: (0.5153072,0.877661) [6]: 8 35 58 65 71 82 
#> 14: (0.09830826,0) [5]: 36 57 61 66 86 
#> 15: (0.5794543,0.4807306) [5]: 16 19 34 46 53 
#> 16: (0.7370465,0.3439962) [10]: 15 19 23 34 45 51 72 73 90 99 
#> 17: (0.1836606,0.2304887) [5]: 24 31 33 36 63 
#> 18: (0.7943964,0.7111173) [6]: 3 6 10 26 34 48 
#> 19: (0.4548386,0.3884795) [6]: 15 16 21 53 73 78 
#> 20: (0.2940803,0.1959671) [6]: 31 33 55 68 85 97 
#> 21: (0.3383256,0.3503165) [6]: 2 19 73 78 85 98 
#> 22: (0.06947969,0.8156346) [5]: 7 28 49 77 81 
#> 23: (0.7877712,0.4944158) [5]: 16 34 48 76 90 
#> 24: (0.1916977,0.2554694) [4]: 11 17 33 63 
#> 25: (0.1117335,0.471513) [5]: 4 12 37 42 80 
#> 26: (0.5742458,0.7057758) [6]: 3 8 18 32 34 43 
#> 27: (0.9825947,0.9276633) [5]: 35 39 56 62 96 
#> 28: (0.131448,0.7749065) [9]: 1 7 9 22 37 42 77 79 93 
#> 29: (0.9854627,0.08954526) [4]: 50 60 74 89 
#> 30: (0.8826108,0.04327769) [5]: 50 72 83 89 94 
#> 31: (0.2258283,0.1548543) [7]: 17 20 33 36 57 59 97 
#> 32: (0.5026046,0.7079632) [5]: 8 26 43 65 95 
#> 33: (0.2383517,0.2608308) [7]: 11 17 20 24 31 85 98 
#> 34: (0.654257,0.5489428) [8]: 15 16 18 23 26 43 46 48 
#> 35: (0.6204141,1) [8]: 13 27 39 54 58 69 71 82 
#> 36: (0.1012567,0.1125712) [6]: 14 17 31 57 63 66 
#> 37: (0.1910354,0.6250357) [6]: 1 4 25 28 42 52 
#> 38: (0.8901494,0.1838986) [5]: 51 74 83 89 99 
#> 39: (0.8589336,0.9446257) [7]: 3 10 27 35 47 71 96 
#> 40: (0.07777007,0.9192984) [4]: 49 54 81 87 
#> 41: (0.03825126,0.6683096) [6]: 12 42 49 63 77 93 
#> 42: (0.1158415,0.6316571) [6]: 12 25 28 37 41 93 
#> 43: (0.5395604,0.6045277) [6]: 26 32 34 46 53 95 
#> 44: (0.5154145,0.07767012) [4]: 55 67 88 97 
#> 45: (0.845916,0.3318166) [4]: 16 51 74 90 
#> 46: (0.6010393,0.5204437) [4]: 15 34 43 53 
#> 47: (0.903265,0.8181644) [5]: 6 10 39 84 96 
#> 48: (0.8136217,0.6814177) [7]: 6 18 23 34 56 62 76 
#> 49: (0,0.8006657) [7]: 22 40 41 63 66 77 81 
#> 50: (0.9091956,0.02407065) [6]: 29 30 86 88 89 94 
#> 51: (0.8518451,0.2872552) [5]: 16 38 45 74 99 
#> 52: (0.2827473,0.5610844) [6]: 1 4 37 64 78 95 
#> 53: (0.5386266,0.5493443) [6]: 15 19 43 46 78 95 
#> 54: (0.2410024,0.9895977) [5]: 35 40 69 70 87 
#> 55: (0.4122456,0.1306076) [5]: 20 44 67 68 97 
#> 56: (1,0.6036345) [5]: 27 48 60 62 76 
#> 57: (0.1446681,0.03337431) [5]: 14 31 36 59 61 
#> 58: (0.4429881,0.9078685) [6]: 5 13 35 65 69 82 
#> 59: (0.2169772,0.02786316) [5]: 31 57 61 86 97 
#> 60: (0.9977757,0.4355533) [5]: 29 56 74 76 90 
#> 61: (0.1599265,0.02288597) [4]: 14 57 59 86 
#> 62: (0.9758555,0.7212268) [6]: 6 27 48 56 84 96 
#> 63: (0.06504696,0.3261832) [9]: 11 12 17 24 36 41 49 66 80 
#> 64: (0.2601581,0.7505434) [5]: 1 9 52 91 95 
#> 65: (0.4536804,0.7744574) [7]: 5 8 13 32 58 91 95 
#> 66: (0.001881215,0.0001720019) [4]: 14 36 49 63 
#> 67: (0.5314487,0.08051148) [6]: 44 55 68 72 88 94 
#> 68: (0.4370106,0.2673845) [6]: 20 55 67 72 73 85 
#> 69: (0.274594,0.9733897) [6]: 5 35 54 58 70 75 
#> 70: (0.2420563,0.9601491) [6]: 54 69 75 87 92 100 
#> 71: (0.6324688,0.9404582) [5]: 3 8 13 35 39 
#> 72: (0.735409,0.1664444) [8]: 16 30 67 68 73 83 94 99 
#> 73: (0.4462513,0.2948664) [6]: 16 19 21 68 72 85 
#> 74: (0.9449086,0.3691798) [7]: 29 38 45 51 60 89 90 
#> 75: (0.2850036,0.8700315) [6]: 5 9 69 70 91 100 
#> 76: (0.9381044,0.5099861) [5]: 23 48 56 60 90 
#> 77: (0.02953472,0.7512234) [5]: 22 28 41 49 93 
#> 78: (0.3823563,0.4665283) [7]: 2 4 19 21 52 53 95 
#> 79: (0.1225114,0.8407039) [5]: 7 9 28 87 92 
#> 80: (0.1401805,0.4163424) [6]: 2 4 11 12 25 63 
#> 81: (0.0636146,0.8311254) [5]: 7 22 40 49 87 
#> 82: (0.475576,0.9073758) [3]: 13 35 58 
#> 83: (0.8781451,0.1449405) [5]: 30 38 72 89 99 
#> 84: (0.9092171,0.8175023) [4]: 6 47 62 96 
#> 85: (0.3376488,0.3198809) [6]: 20 21 33 68 73 98 
#> 86: (0.2512925,0.001279666) [6]: 14 50 59 61 88 97 
#> 87: (0.08914897,0.9143498) [7]: 7 40 54 70 79 81 92 
#> 88: (0.5660539,0.03269834) [6]: 44 50 67 86 94 97 
#> 89: (0.958219,0.1172698) [6]: 29 30 38 50 74 83 
#> 90: (0.8754012,0.4109329) [6]: 16 23 45 60 74 76 
#> 91: (0.2881229,0.7832835) [6]: 5 9 64 65 75 95 
#> 92: (0.1623021,0.8669802) [5]: 9 70 79 87 100 
#> 93: (0.04584944,0.7169532) [4]: 28 41 42 77 
#> 94: (0.6768708,0.08109197) [5]: 30 50 67 72 88 
#> 95: (0.4708485,0.6257525) [8]: 32 43 52 53 64 65 78 91 
#> 96: (0.9330483,0.83448) [5]: 27 39 47 62 84 
#> 97: (0.3661773,0.08720362) [7]: 20 31 44 55 59 86 88 
#> 98: (0.2660567,0.3366931) [5]: 2 11 21 33 85 
#> 99: (0.7975027,0.2191768) [5]: 16 38 51 72 83 
#> 100: (0.2213311,0.8909305) [4]: 9 70 75 92 
#> number of nodes: 100 
#> number of arcs: 285 
#> number of boundary nodes: 12 
#> boundary nodes:  14 27 29 35 40 49 50 54 56 60 66 86 
#> number of triangles: 186 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.083565912 0.161282044 0.111807216 0.190800902 0.078730870 0.082735288 
#>                                                                         
#> 0.094613340 0.073541769 0.124273474 0.174801656 0.093658279 0.226503160 
#>                                                                         
#> 0.151249327 0.092526002 0.220215421 0.116822950 0.125464975 0.137341853 
#>                                                                         
#> 0.172794774 0.062072767 0.165792637 0.133838862 0.180691984 0.103881045 
#>                                                                         
#> 0.086804081 0.117088218 0.005988857 0.079663046 0.035379020 0.167046652 
#>                                                                         
#> 0.080850307 0.087053449 0.016563906 0.074154247 0.066401515 0.132929739 
#>                                                                         
#> 0.120203362 0.082553419 0.071674565 0.043057434 0.067014509 0.063702781 
#>                                                                         
#> 0.047683773 0.134006392 0.145054643 0.046961061 0.080762920 0.117361526 
#>                                                                         
#> 0.085811746 0.343174161 0.160196826 0.060749841 0.171922480 0.032591586 
#>                                                                         
#> 0.093885984 0.018515527 0.090314465 0.150030242 0.068779245 0.181377829 
#>                                                                         
#> 0.285692263 0.221036684 0.060368260 0.316392791 0.081463730 0.087107810 
#>                                                                         
#> 0.044954131 0.084432009 0.121030636 0.144219414 0.094006146 0.216659165 
#>                                                                         
#> 0.131552042 0.106713972 0.176063507 0.207235873 0.176858405 0.121336932 
#>                                                                         
#> 0.112377188 0.115528446 0.155802564 0.063294492 0.139000828 0.073539648 
#>                                                                         
#> 0.111446015 0.070530886 0.057592041 0.104628866 0.117231465 0.211860213 
#>                                                                         
#> 0.075484847 0.109839066 0.127455833 0.244862930 0.132759374 0.121073363 
#>                                                                         
#> 0.120045371 0.037955433 0.110337323 0.084855012 0.252261368 0.105306140 
#>                                                                         
#> 0.239220078 0.084720145 0.144346864 0.103497584 0.072518793 0.160567615 
#>                                                                         
#> 0.149692708 0.071915277 0.104162230 0.037297456 0.332074065 0.123985679 
#>                                                                         
#> 0.109609431 0.033959406 0.169474803 0.478920489 0.102137516 0.207175013 
#>                                                                         
#> 0.059022157 0.129301961 0.081083905 0.029260328 0.112254597 0.120922836 
#>                                                                         
#> 0.316325551 0.182162583 0.244873898 0.159618988 0.035128606 0.209380207 
#>                                                                         
#> 0.095398731 0.057267357 0.143462306 0.241256503 0.315008732 0.099828106 
#>                                                                         
#> 0.087150739 0.115500584 0.115815769 0.175476707 0.029520130 0.084137089 
#>                                                                         
#> 0.086940167 0.100473361 0.147188795 0.180827761 0.062114803 0.067834393 
#>                                                                         
#> 0.152885389 0.162945188 0.209414921 0.132790866 0.083398953 0.162734121 
#>                                                                         
#> 0.109606313 0.145254354 0.096849534 0.148721315 0.123001827 0.128330742 
#>                                                                         
#> 0.047163972 0.085868197 0.091624220 0.089477422 0.095841421 0.075017654 
#>                                                                         
#> 0.075781243 0.112393808 0.085486847 0.030625718 0.028078155 0.065704540 
#>                                                                         
#> 0.030661625 0.103656879 0.130106819 0.119808372 0.084545462 0.098403265 
#>                                                                         
#> 0.082860646 0.073444903 0.090777733 0.086679658 0.103441284 0.107308703 
#>                                                                         
#> 0.198691798 0.145729782 0.090956496 0.043586899 0.166743530 0.161289462 
#>                                                                         
#> 0.049613916 0.078374402 0.065731075 0.057123352 0.112609849 0.096427199 
#>                                                                         
#> 0.079841999 0.155046292 0.208642217 0.101234121 0.045200008 0.177559406 
#>                                                                         
#> 0.138689700 0.128055745 0.109548745 0.153696327 0.158742076 0.294916247 
#>                                                                         
#> 0.143857275 0.086594932 0.062544152 0.026241814 0.152402822 0.214335310 
#>                                                                         
#> 0.106514317 0.122603753 0.131350080 0.085516002 0.079678131 0.130489294 
#>                                                                         
#> 0.135036638 0.159779532 0.030443109 0.071073865 0.075791831 0.151137371 
#>                                                                         
#> 0.188780225 0.107024527 0.124818983 0.105536589 0.206546518 0.324495925 
#>                                                                         
#> 0.103371592 0.144096970 0.282559912 0.038869903 0.100516653 0.032797249 
#>                                                                         
#> 0.105789585 0.101760886 0.191933371 0.209186249 0.146097084 0.127299179 
#>                                                                         
#> 0.088130846 0.346842407 0.199920507 0.193203800 0.110226198 0.099136135 
#>                                                                         
#> 0.040765552 0.095251573 0.089302104 0.012408414 0.177726678 0.083370749 
#>                                                                         
#> 0.137772679 0.049233503 0.055191303 0.149541461 0.067725956 0.016283982 
#>                                                                         
#> 0.115957752 0.105808944 0.201958180 0.343250168 0.658297737 0.198907067 
#>                                                                         
#> 0.029467473 0.168095964 0.043407658 0.124826623 0.145423248 0.221323862 
#>                                                                         
#> 0.028993908 0.122642446 0.072254746 0.152989584 0.369333631 0.346227018 
#>                                     
#> 0.379554343 0.141851675 0.800495956 
#> 
#> $alpha
#> [1] 0.2664714
#> 
#> attr(,"class")
#> [1] "scree"