Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1         y1           x2         y2          mx1
#>    32    4 0.9832767746 0.33088983 0.9013516232 0.42975245  0.920834332
#>     1    4 0.9320831406 0.29343386 0.9013516232 0.42975245  0.853941065
#>    86    4 0.7857303443 0.41354876 0.9013516232 0.42975245  0.837236741
#>    30   78 0.9910335084 0.53097616 0.8890089795 0.51946094  0.944960624
#>    32    1 0.9832767746 0.33088983 0.9320831406 0.29343386  1.007846828
#>    14    1 0.9327458737 0.19276369 0.9320831406 0.29343386  0.774518762
#>    32   30 0.9832767746 0.33088983 0.9910335084 0.53097616  1.002664170
#>    17   77 0.7360947736 0.42029672 0.6592149878 0.36110097  0.748389534
#>    86   22 0.7857303443 0.41354876 0.8366833721 0.55035885  0.819600024
#>    78   22 0.8890089795 0.51946094 0.8366833721 0.55035885  0.827903931
#>    11   22 0.8376727524 0.59882759 0.8366833721 0.55035885  0.885694116
#>    17   86 0.7360947736 0.42029672 0.7857303443 0.41354876  0.767793961
#>    44   22 0.7616426505 0.52409888 0.8366833721 0.55035885  0.785720036
#>    30   99 0.9910335084 0.53097616 0.9160673498 0.66064019  0.933370812
#>     1   86 0.9320831406 0.29343386 0.7857303443 0.41354876  0.777578500
#>    19   77 0.6609343789 0.40184143 0.6592149878 0.36110097  0.706260817
#>    44   89 0.7616426505 0.52409888 0.6792979125 0.50213016  0.730157834
#>    17   89 0.7360947736 0.42029672 0.6792979125 0.50213016  0.699542712
#>    13   16 0.6650703676 0.49284798 0.6388208140 0.41177121  0.667979209
#>    19   17 0.6609343789 0.40184143 0.7360947736 0.42029672  0.689907507
#>    10   92 0.7237498023 0.60603088 0.7018232809 0.56594370  0.747216142
#>    93   11 0.7563972900 0.66140742 0.8376727524 0.59882759  0.822313832
#>    99   11 0.9160673498 0.66064019 0.8376727524 0.59882759  0.909160314
#>    66   13 0.6187003792 0.50205670 0.6650703676 0.49284798  0.648961578
#>    11   10 0.8376727524 0.59882759 0.7237498023 0.60603088  0.779422746
#>     9   26 0.7960683717 0.04017765 0.8762108796 0.04398199  0.831196168
#>    92   66 0.7018232809 0.56594370 0.6187003792 0.50205670  0.648942614
#>    77   56 0.6592149878 0.36110097 0.6767883619 0.09303242  0.769382089
#>    39    9 0.7306851647 0.05858514 0.7960683717 0.04017765  0.798807488
#>    16   64 0.6388208140 0.41177121 0.5695686466 0.44994199  0.576080803
#>    77   16 0.6592149878 0.36110097 0.6388208140 0.41177121  0.588743621
#>    16   66 0.6388208140 0.41177121 0.6187003792 0.50205670  0.617128330
#>    93   74 0.7563972900 0.66140742 0.8037930344 0.72622822  0.773183083
#>    56   39 0.6767883619 0.09303242 0.7306851647 0.05858514  0.782742766
#>    24   74 0.9099384676 0.72737438 0.8037930344 0.72622822  0.857275124
#>    29   24 1.0000000000 0.73513725 0.9099384676 0.72737438  0.957825057
#>    30   29 0.9910335084 0.53097616 1.0000000000 0.73513725  1.016391073
#>    93   34 0.7563972900 0.66140742 0.7992386080 0.73059644  0.733005378
#>    24   90 0.9099384676 0.72737438 0.8858693880 0.80721824  0.856563055
#>    10   93 0.7237498023 0.60603088 0.7563972900 0.66140742  0.700544780
#>    29   96 1.0000000000 0.73513725 0.9747541543 0.78986000  0.953618669
#>     7    2 0.6102308867 0.12076998 0.6858758383 0.00000000  0.615671508
#>     7   77 0.6102308867 0.12076998 0.6592149878 0.36110097  0.615443399
#>    64   66 0.5695686466 0.44994199 0.6187003792 0.50205670  0.555987781
#>    47   92 0.5526730694 0.57536616 0.7018232809 0.56594370  0.629362872
#>    81   80 0.4226263116 0.52134545 0.4440672753 0.45838370  0.502205932
#>    96   43 0.9747541543 0.78986000 0.9257942391 0.80238815  0.941026919
#>    24   43 0.9099384676 0.72737438 0.9257942391 0.80238815  0.901422567
#>     2   39 0.6858758383 0.00000000 0.7306851647 0.05858514  0.685260116
#>    34   90 0.7992386080 0.73059644 0.8858693880 0.80721824  0.839034254
#>    64   47 0.5695686466 0.44994199 0.5526730694 0.57536616  0.510291747
#>    47   95 0.5526730694 0.57536616 0.6509760081 0.68313210  0.595262790
#>    43   82 0.9257942391 0.80238815 0.9899235280 0.80538069  0.953326911
#>    34   68 0.7992386080 0.73059644 0.7945616719 0.81022803  0.746981575
#>    95   93 0.6509760081 0.68313210 0.7563972900 0.66140742  0.711193504
#>   100   68 0.7657701867 0.84826841 0.7945616719 0.81022803  0.828048645
#>    46   50 0.5067127083 0.64085980 0.5373978378 0.58996027  0.581919941
#>    41   95 0.6464687277 0.73825932 0.6509760081 0.68313210  0.695762430
#>    53   35 0.4738360077 0.18617326 0.5083524714 0.09139570  0.544278373
#>     7    5 0.6102308867 0.12076998 0.5256790297 0.33070852  0.578776725
#>    38   55 0.4211508341 0.46857894 0.3797033383 0.41911477  0.417506854
#>     5   64 0.5256790297 0.33070852 0.5695686466 0.44994199  0.503613676
#>    35    2 0.5083524714 0.09139570 0.6858758383 0.00000000  0.584029916
#>    41   25 0.6464687277 0.73825932 0.5024467113 0.67934023  0.576110837
#>    25   46 0.5024467113 0.67934023 0.5067127083 0.64085980  0.577045858
#>    42   73 0.6830194157 0.76773958 0.6626284435 0.80045479  0.650755917
#>    42  100 0.6830194157 0.76773958 0.7657701867 0.84826841  0.734121396
#>    90   43 0.8858693880 0.80721824 0.9257942391 0.80238815  0.914720297
#>    21    5 0.3950913064 0.31692778 0.5256790297 0.33070852  0.454940465
#>    50   81 0.5373978378 0.58996027 0.4226263116 0.52134545  0.495106674
#>    73   98 0.6626284435 0.80045479 0.6813494192 0.82526283  0.645250189
#>    81   38 0.4226263116 0.52134545 0.4211508341 0.46857894  0.446310690
#>    81   46 0.4226263116 0.52134545 0.5067127083 0.64085980  0.431555418
#>    94   25 0.5994768346 0.78070193 0.5024467113 0.67934023  0.575312092
#>    73   94 0.6626284435 0.80045479 0.5994768346 0.78070193  0.636199109
#>    53   79 0.4738360077 0.18617326 0.2955681278 0.18335051  0.385575527
#>    53    5 0.4738360077 0.18617326 0.5256790297 0.33070852  0.466006814
#>    55   21 0.3797033383 0.41911477 0.3950913064 0.31692778  0.449345552
#>    83   62 0.5931613656 0.78579195 0.5455459787 0.78937899  0.566212780
#>   100   76 0.7657701867 0.84826841 0.8537390501 0.92477783  0.798425309
#>    98  100 0.6813494192 0.82526283 0.7657701867 0.84826841  0.713416408
#>    68   90 0.7945616719 0.81022803 0.8858693880 0.80721824  0.841851660
#>    62   51 0.5455459787 0.78937899 0.5167828198 0.76634629  0.564632150
#>    51   25 0.5167828198 0.76634629 0.5024467113 0.67934023  0.568615674
#>    83   94 0.5931613656 0.78579195 0.5994768346 0.78070193  0.621527659
#>    51   37 0.5167828198 0.76634629 0.4333691279 0.76703888  0.474759576
#>     6   33 0.3478301445 0.40523927 0.2864820551 0.37508600  0.339781647
#>    62   65 0.5455459787 0.78937899 0.6151227780 0.85438921  0.555400489
#>    65   98 0.6151227780 0.85438921 0.6813494192 0.82526283  0.645404394
#>    48   98 0.6361404762 0.86905576 0.6813494192 0.82526283  0.695511165
#>    81   69 0.4226263116 0.52134545 0.4127326362 0.68578608  0.365733708
#>    37   69 0.4333691279 0.76703888 0.4127326362 0.68578608  0.460064763
#>    72    6 0.3335945662 0.50846293 0.3478301445 0.40523927  0.344114330
#>    48   23 0.6361404762 0.86905576 0.7429031572 0.92431579  0.665854592
#>    23   76 0.7429031572 0.92431579 0.8537390501 0.92477783  0.798325415
#>    76   82 0.8537390501 0.92477783 0.9899235280 0.80538069  1.308867418
#>    27   79 0.3035614852 0.27393735 0.2955681278 0.18335051  0.384125376
#>    33   27 0.2864820551 0.37508600 0.3035614852 0.27393735  0.332681537
#>    15   72 0.2491474388 0.55183000 0.3335945662 0.50846293  0.323355285
#>    63   65 0.4958147194 0.84148725 0.6151227780 0.85438921  0.549598912
#>    72   33 0.3335945662 0.50846293 0.2864820551 0.37508600  0.287954154
#>    23   58 0.7429031572 0.92431579 0.8150407644 0.97636019  0.718624163
#>    21   27 0.3950913064 0.31692778 0.3035614852 0.27393735  0.384164848
#>    37   63 0.4333691279 0.76703888 0.4958147194 0.84148725  0.457158698
#>    65   48 0.6151227780 0.85438921 0.6361404762 0.86905576  0.580503963
#>    60    2 0.2294124287 0.07324186 0.6858758383 0.00000000  0.396611668
#>    60   79 0.2294124287 0.07324186 0.2955681278 0.18335051  0.191318888
#>    72   69 0.3335945662 0.50846293 0.4127326362 0.68578608  0.358596921
#>    52   91 0.3152458324 0.78217429 0.2832531542 0.66690546  0.342724788
#>    37   88 0.4333691279 0.76703888 0.4773776941 0.85572863  0.432083547
#>    49   85 0.4754280624 0.93507666 0.4576064067 0.87269189  0.494638743
#>    15   87 0.2491474388 0.55183000 0.2164850756 0.39571346  0.259548136
#>    87   33 0.2164850756 0.39571346 0.2864820551 0.37508600  0.272050904
#>    37   52 0.4333691279 0.76703888 0.3152458324 0.78217429  0.369863141
#>    45   65 0.5444747351 0.99892072 0.6151227780 0.85438921  0.579195550
#>    88   63 0.4773776941 0.85572863 0.4958147194 0.84148725  0.523638529
#>    57   36 0.2061666000 0.64019830 0.1163884527 0.58455432  0.184445217
#>     8   52 0.3447138261 0.86658991 0.3152458324 0.78217429  0.378514441
#>    49   65 0.4754280624 0.93507666 0.6151227780 0.85438921  0.557385507
#>    28   60 0.2841240813 0.10317732 0.2294124287 0.07324186  0.383583269
#>    28   79 0.2841240813 0.10317732 0.2955681278 0.18335051  0.221291089
#>    59   87 0.1499548879 0.27512420 0.2164850756 0.39571346  0.160109223
#>    15   91 0.2491474388 0.55183000 0.2832531542 0.66690546  0.259310296
#>    85    8 0.4576064067 0.87269189 0.3447138261 0.86658991  0.403327776
#>    49   45 0.4754280624 0.93507666 0.5444747351 0.99892072  0.349228602
#>    84   57 0.1820719555 0.65596198 0.2061666000 0.64019830  0.228434019
#>    52   61 0.3152458324 0.78217429 0.2604417267 0.76460365  0.303920876
#>    27   87 0.3035614852 0.27393735 0.2164850756 0.39571346  0.227041836
#>    36   20 0.1163884527 0.58455432 0.1735863674 0.49164631  0.180782440
#>    61   84 0.2604417267 0.76460365 0.1820719555 0.65596198  0.227897261
#>    97   67 0.1208218221 0.28575779 0.0986972045 0.12165868  0.107492401
#>    12   97 0.0643051965 0.45172619 0.1208218221 0.28575779  0.124113933
#>     3   36 0.0728951147 0.79090864 0.1163884527 0.58455432  0.079374950
#>    79   59 0.2955681278 0.18335051 0.1499548879 0.27512420  0.189583544
#>    20   12 0.1735863674 0.49164631 0.0643051965 0.45172619  0.138424042
#>    75   12 0.0221526268 0.28716866 0.0643051965 0.45172619  0.057315793
#>    40   12 0.0000000000 0.43018464 0.0643051965 0.45172619  0.007164683
#>    60   67 0.2294124287 0.07324186 0.0986972045 0.12165868  0.078147508
#>    67   75 0.0986972045 0.12165868 0.0221526268 0.28716866  0.055891003
#>    75   40 0.0221526268 0.28716866 0.0000000000 0.43018464 -1.080977203
#>     3   18 0.0728951147 0.79090864 0.0004654966 0.60095039  0.071160766
#>    61    3 0.2604417267 0.76460365 0.0728951147 0.79090864  0.163089869
#>    61    8 0.2604417267 0.76460365 0.3447138261 0.86658991  0.223235111
#>    12   36 0.0643051965 0.45172619 0.1163884527 0.58455432  0.050105653
#>     8   49 0.3447138261 0.86658991 0.4754280624 0.93507666  0.333487329
#>     3   31 0.0728951147 0.79090864 0.1627892487 0.98753316  0.164011685
#>    78    4 0.8890089795 0.51946094 0.9013516232 0.42975245  0.837236741
#>    77    1 0.6592149878 0.36110097 0.9320831406 0.29343386  0.774518762
#>     4   30 0.9013516232 0.42975245 0.9910335084 0.53097616  1.002664170
#>    86   77 0.7857303443 0.41354876 0.6592149878 0.36110097  0.748389534
#>    44   86 0.7616426505 0.52409888 0.7857303443 0.41354876  0.767793961
#>    78   99 0.8890089795 0.51946094 0.9160673498 0.66064019  0.933370812
#>    13   89 0.6650703676 0.49284798 0.6792979125 0.50213016  0.699542712
#>    92   44 0.7018232809 0.56594370 0.7616426505 0.52409888  0.717417614
#>    16   19 0.6388208140 0.41177121 0.6609343789 0.40184143  0.638902460
#>    13   17 0.6650703676 0.49284798 0.7360947736 0.42029672  0.689907507
#>    74   11 0.8037930344 0.72622822 0.8376727524 0.59882759  0.822313832
#>    78   11 0.8890089795 0.51946094 0.8376727524 0.59882759  0.909160314
#>    44   10 0.7616426505 0.52409888 0.7237498023 0.60603088  0.779422746
#>    14   26 0.9327458737 0.19276369 0.8762108796 0.04398199  0.831196168
#>    89   66 0.6792979125 0.50213016 0.6187003792 0.50205670  0.648942614
#>    95   10 0.6509760081 0.68313210 0.7237498023 0.60603088  0.657387806
#>    14   56 0.9327458737 0.19276369 0.6767883619 0.09303242  0.769382089
#>    14    9 0.9327458737 0.19276369 0.7960683717 0.04017765  0.798807488
#>     5   16 0.5256790297 0.33070852 0.6388208140 0.41177121  0.588743621
#>    34   74 0.7992386080 0.73059644 0.8037930344 0.72622822  0.773183083
#>    14   39 0.9327458737 0.19276369 0.7306851647 0.05858514  0.782742766
#>    99   74 0.9160673498 0.66064019 0.8037930344 0.72622822  0.857275124
#>    99   24 0.9160673498 0.66064019 0.9099384676 0.72737438  0.957825057
#>    99   29 0.9160673498 0.66064019 1.0000000000 0.73513725  1.016391073
#>    29   32 1.0000000000 0.73513725 0.9832767746 0.33088983 40.292239323
#>    42   34 0.6830194157 0.76773958 0.7992386080 0.73059644  0.733005378
#>    74   90 0.8037930344 0.72622822 0.8858693880 0.80721824  0.856563055
#>    24   96 0.9099384676 0.72737438 0.9747541543 0.78986000  0.953618669
#>    56    2 0.6767883619 0.09303242 0.6858758383 0.00000000  0.615671508
#>     5   77 0.5256790297 0.33070852 0.6592149878 0.36110097  0.615443399
#>    47   66 0.5526730694 0.57536616 0.6187003792 0.50205670  0.555987781
#>    95   92 0.6509760081 0.68313210 0.7018232809 0.56594370  0.629362872
#>    47   80 0.5526730694 0.57536616 0.4440672753 0.45838370  0.502205932
#>     2    9 0.6858758383 0.00000000 0.7960683717 0.04017765  0.851584350
#>    50   95 0.5373978378 0.58996027 0.6509760081 0.68313210  0.595262790
#>    82   96 0.9899235280 0.80538069 0.9747541543 0.78986000  0.956995326
#>    82   29 0.9899235280 0.80538069 1.0000000000 0.73513725  1.008366453
#>    42   68 0.6830194157 0.76773958 0.7945616719 0.81022803  0.746981575
#>    42   93 0.6830194157 0.76773958 0.7563972900 0.66140742  0.711193504
#>    76   68 0.8537390501 0.92477783 0.7945616719 0.81022803  0.828048645
#>    42   95 0.6830194157 0.76773958 0.6509760081 0.68313210  0.695762430
#>     7   35 0.6102308867 0.12076998 0.5083524714 0.09139570  0.544278373
#>    80   55 0.4440672753 0.45838370 0.3797033383 0.41911477  0.417506854
#>    80   64 0.4440672753 0.45838370 0.5695686466 0.44994199  0.503613676
#>    95   25 0.6509760081 0.68313210 0.5024467113 0.67934023  0.576110837
#>    95   46 0.6509760081 0.68313210 0.5067127083 0.64085980  0.577045858
#>    41   73 0.6464687277 0.73825932 0.6626284435 0.80045479  0.650755917
#>    76   43 0.8537390501 0.92477783 0.9257942391 0.80238815  0.914720297
#>    80    5 0.4440672753 0.45838370 0.5256790297 0.33070852  0.454940465
#>    47   81 0.5526730694 0.57536616 0.4226263116 0.52134545  0.495106674
#>    98   42 0.6813494192 0.82526283 0.6830194157 0.76773958  0.693239839
#>    80   38 0.4440672753 0.45838370 0.4211508341 0.46857894  0.446310690
#>    69   46 0.4127326362 0.68578608 0.5067127083 0.64085980  0.431555418
#>    41   94 0.6464687277 0.73825932 0.5994768346 0.78070193  0.636199109
#>    80   21 0.4440672753 0.45838370 0.3950913064 0.31692778  0.449345552
#>    94   62 0.5994768346 0.78070193 0.5455459787 0.78937899  0.566212780
#>    65   73 0.6151227780 0.85438921 0.6626284435 0.80045479  0.623707389
#>    72   81 0.3335945662 0.50846293 0.4226263116 0.52134545  0.380829837
#>    23  100 0.7429031572 0.92431579 0.7657701867 0.84826841  0.713416408
#>    76   90 0.8537390501 0.92477783 0.8858693880 0.80721824  0.841851660
#>    94   51 0.5994768346 0.78070193 0.5167828198 0.76634629  0.564632150
#>    65   94 0.6151227780 0.85438921 0.5994768346 0.78070193  0.621527659
#>    25   37 0.5024467113 0.67934023 0.4333691279 0.76703888  0.474759576
#>    69   25 0.4127326362 0.68578608 0.5024467113 0.67934023  0.455585472
#>    21   33 0.3950913064 0.31692778 0.2864820551 0.37508600  0.339781647
#>     6   55 0.3478301445 0.40523927 0.3797033383 0.41911477  0.383260452
#>    72   38 0.3335945662 0.50846293 0.4211508341 0.46857894  0.369013178
#>    23   98 0.7429031572 0.92431579 0.6813494192 0.82526283  0.695511165
#>    71   23 0.6115644243 1.00000000 0.7429031572 0.92431579  0.665854592
#>    58   76 0.8150407644 0.97636019 0.8537390501 0.92477783  0.798325415
#>    58   82 0.8150407644 0.97636019 0.9899235280 0.80538069  1.308867418
#>    91   72 0.2832531542 0.66690546 0.3335945662 0.50846293  0.323355285
#>    63   62 0.4958147194 0.84148725 0.5455459787 0.78937899  0.509570032
#>    63   51 0.4958147194 0.84148725 0.5167828198 0.76634629  0.475313256
#>    71   58 0.6115644243 1.00000000 0.8150407644 0.97636019  0.718624163
#>    28   53 0.2841240813 0.10317732 0.4738360077 0.18617326  0.396610732
#>    53   27 0.4738360077 0.18617326 0.3035614852 0.27393735  0.384164848
#>    59   27 0.1499548879 0.27512420 0.3035614852 0.27393735  0.226453487
#>    71   48 0.6115644243 1.00000000 0.6361404762 0.86905576  0.580503963
#>    67   79 0.0986972045 0.12165868 0.2955681278 0.18335051  0.191318888
#>    91   69 0.2832531542 0.66690546 0.4127326362 0.68578608  0.358596921
#>    85   88 0.4576064067 0.87269189 0.4773776941 0.85572863  0.432083547
#>    72   87 0.3335945662 0.50846293 0.2164850756 0.39571346  0.259548136
#>    69   52 0.4127326362 0.68578608 0.3152458324 0.78217429  0.369863141
#>    71   65 0.6115644243 1.00000000 0.6151227780 0.85438921  0.579195550
#>    49   63 0.4754280624 0.93507666 0.4958147194 0.84148725  0.523638529
#>    15   36 0.2491474388 0.55183000 0.1163884527 0.58455432  0.184445217
#>    35   60 0.5083524714 0.09139570 0.2294124287 0.07324186  0.383583269
#>    97   87 0.1208218221 0.28575779 0.2164850756 0.39571346  0.160109223
#>    20   15 0.1735863674 0.49164631 0.2491474388 0.55183000  0.252922162
#>    57   91 0.2061666000 0.64019830 0.2832531542 0.66690546  0.259310296
#>    37    8 0.4333691279 0.76703888 0.3447138261 0.86658991  0.403327776
#>    31   45 0.1627892487 0.98753316 0.5444747351 0.99892072  0.349228602
#>    91   61 0.2832531542 0.66690546 0.2604417267 0.76460365  0.303920876
#>    36   84 0.1163884527 0.58455432 0.1820719555 0.65596198  0.165890772
#>    91   84 0.2832531542 0.66690546 0.1820719555 0.65596198  0.227897261
#>    59   67 0.1499548879 0.27512420 0.0986972045 0.12165868  0.107492401
#>    87   12 0.2164850756 0.39571346 0.0643051965 0.45172619  0.138424042
#>    75   97 0.0221526268 0.28716866 0.1208218221 0.28575779  0.070379827
#>    18   12 0.0004654966 0.60095039 0.0643051965 0.45172619  0.007164683
#>    54   67 0.0240708916 0.11672718 0.0986972045 0.12165868  0.078147508
#>    54   75 0.0240708916 0.11672718 0.0221526268 0.28716866  0.055891003
#>    54   40 0.0240708916 0.11672718 0.0000000000 0.43018464 -1.080977203
#>    36   18 0.1163884527 0.58455432 0.0004654966 0.60095039  0.071160766
#>    84    3 0.1820719555 0.65596198 0.0728951147 0.79090864  0.163089869
#>    31   61 0.1627892487 0.98753316 0.2604417267 0.76460365  0.178417913
#>    31    8 0.1627892487 0.98753316 0.3447138261 0.86658991  0.223235111
#>    31   49 0.1627892487 0.98753316 0.4754280624 0.93507666  0.333487329
#>    70    3 0.0851480491 0.95779693 0.0728951147 0.79090864 -0.552953627
#>    70   31 0.0851480491 0.95779693 0.1627892487 0.98753316  0.164011685
#>    78   86 0.8890089795 0.51946094 0.7857303443 0.41354876  0.837236741
#>    14   32 0.9327458737 0.19276369 0.9832767746 0.33088983  1.007846828
#>    77   14 0.6592149878 0.36110097 0.9327458737 0.19276369  0.774518762
#>    44   17 0.7616426505 0.52409888 0.7360947736 0.42029672  0.767793961
#>    11   44 0.8376727524 0.59882759 0.7616426505 0.52409888  0.785720036
#>    92   89 0.7018232809 0.56594370 0.6792979125 0.50213016  0.717417614
#>    19   13 0.6609343789 0.40184143 0.6650703676 0.49284798  0.667979209
#>    56    7 0.6767883619 0.09303242 0.6102308867 0.12076998  0.615671508
#>     2   26 0.6858758383 0.00000000 0.8762108796 0.04398199  0.851584350
#>    50   47 0.5373978378 0.58996027 0.5526730694 0.57536616  0.595262790
#>    42   41 0.6830194157 0.76773958 0.6464687277 0.73825932  0.695762430
#>     7   53 0.6102308867 0.12076998 0.4738360077 0.18617326  0.544278373
#>    21   53 0.3950913064 0.31692778 0.4738360077 0.18617326  0.466006814
#>    65   83 0.6151227780 0.85438921 0.5931613656 0.78579195  0.621527659
#>    21    6 0.3950913064 0.31692778 0.3478301445 0.40523927  0.339781647
#>    72   55 0.3335945662 0.50846293 0.3797033383 0.41911477  0.369013178
#>    28   35 0.2841240813 0.10317732 0.5083524714 0.09139570  0.396610732
#>    85   37 0.4576064067 0.87269189 0.4333691279 0.76703888  0.432083547
#>    88   49 0.4773776941 0.85572863 0.4754280624 0.93507666  0.494638743
#>    71   45 0.6115644243 1.00000000 0.5444747351 0.99892072  0.579195550
#>    15   57 0.2491474388 0.55183000 0.2061666000 0.64019830  0.184445217
#>    97   59 0.1208218221 0.28575779 0.1499548879 0.27512420  0.160109223
#>    20   87 0.1735863674 0.49164631 0.2164850756 0.39571346  0.252922162
#>    18   40 0.0004654966 0.60095039 0.0000000000 0.43018464  0.007164683
#>    54   60 0.0240708916 0.11672718 0.2294124287 0.07324186  0.078147508
#>    70   18 0.0851480491 0.95779693 0.0004654966 0.60095039 -0.552953627
#>          my1         mx2          my2 bp1 bp2
#>   0.36252128  1.00266417  0.430331757   0   0
#>   0.34744093  0.92083433  0.362521275   0   0
#>   0.46663446  0.85394106  0.347440929   0   0
#>   0.48145578  0.93337081  0.584141192   0   0
#>   0.24359536  0.92083433  0.362521275   0   0
#>   0.24205931  1.00784683  0.243595360   0   0
#>   0.43033176 40.29223932 -1.092804668   0   0
#>   0.32480780  0.70626082  0.379521984   0   0
#>   0.47882789  0.82790393  0.475735212   0   0
#>   0.47573521  0.88569412  0.573602877   0   0
#>   0.57360288  0.78572004  0.575643622   0   0
#>   0.46753990  0.74838953  0.324807801   0   0
#>   0.57564362  0.81960002  0.478827887   0   0
#>   0.58414119  1.01639107  0.632139929   0   0
#>   0.25439773  0.85394106  0.347440929   0   0
#>   0.37952198  0.63890246  0.382364744   0   0
#>   0.47680293  0.71741761  0.524556746   0   0
#>   0.45555437  0.73015783  0.476802933   0   0
#>   0.44711852  0.63406973  0.458097113   0   0
#>   0.44612194  0.70626082  0.379521984   0   0
#>   0.56715530  0.65738781  0.616288792   0   0
#>   0.66294833  0.78115820  0.609497509   0   0
#>   0.58878137  0.84576406  0.669184457   0   0
#>   0.53308417  0.63406973  0.458097113   0   0
#>   0.58205058  0.78115820  0.609497509   0   0
#>   0.14621907  0.85158435 -0.283279909   0   0
#>   0.54872755  0.62761589  0.576475588   0   0
#>   0.23371274  0.69430677  0.228791140   0   0
#>   0.17523092  0.74896123 -0.001822419   0   0
#>   0.37985030  0.61712833  0.454321663   0   0
#>   0.36217640  0.63890246  0.382364744   0   0
#>   0.45432166  0.63406973  0.458097113   0   0
#>   0.69887180  0.82231383  0.662948327   0   0
#>   0.19942294  0.68526012  0.046899902   0   0
#>   0.68888923  0.85656305  0.754834028   0   0
#>   0.69812376  0.95361867  0.746924489   0   0
#>   0.63213993 40.29223932 -1.092804668   0   0
#>   0.72374954  0.77318308  0.698871803   0   0
#>   0.75483403  0.90142257  0.768357010   0   0
#>   0.65702351  0.78115820  0.609497509   0   0
#>   0.74692449  1.00836645  0.772181882   0   0
#>   0.04010243  0.58402992  0.020283547   0   0
#>   0.24486502  0.69430677  0.228791140   0   0
#>   0.51196261  0.61712833  0.454321663   0   0
#>   0.60412896  0.62761589  0.576475588   0   0
#>   0.51331383  0.44631069  0.494279294   0   0
#>   0.75998578  0.95699533  0.822390201   0   0
#>   0.76835701  0.94102692  0.759985780   0   0
#>   0.04689990  0.74896123 -0.001822419   0   0
#>   0.77288686  0.84149910  0.770100039   0   0
#>   0.50580701  0.55598778  0.511962610   0   0
#>   0.63523469  0.62936287  0.604128958   0   0
#>   0.90100319  0.95699533  0.822390201   0   0
#>   0.76748041  0.83903425  0.772886862   0   0
#>   0.70869756  0.70054478  0.657023513   0   0
#>   0.86548904  0.73614463  0.795929969   0   0
#>   0.65149986  0.46488995  0.580947530   0   0
#>   0.71454178  0.57611084  0.704758891   0   0
#>   0.15815329  0.39661073  0.104375104   0   0
#>   0.23009767  0.61544340  0.244865019   0   0
#>   0.42953521  0.36901318  0.470169499   0   0
#>   0.40652531  0.57608080  0.379850301   0   0
#>   0.02028355  0.39661167 -0.343750186   0   0
#>   0.70475889  0.57531209  0.706711345   0   0
#>   0.66813372  0.45558547  0.654668442   0   0
#>   0.77034247  0.69323984  0.796822164   0   0
#>   0.79800902  0.73614463  0.795929969   0   0
#>   0.87827410  0.90142257  0.768357010   0   0
#>   0.37541272  0.46600681  0.270546852   0   0
#>   0.53040423  0.46488995  0.580947530   0   0
#>   0.83303676  0.69323984  0.796822164   0   0
#>   0.49427929  0.38082984  0.496110294   0   0
#>   0.60440062  0.46488995  0.580947530   0   0
#>   0.70671135  0.56861567  0.713121600   0   0
#>   0.77412464  0.62370739  0.814061769   0   0
#>   0.12959935  0.38412538  0.221182330   0   0
#>   0.27054685  0.57877673  0.230097671   0   0
#>   0.37734983  0.38326045  0.367398318   0   0
#>   0.74589238  0.57256400  0.830200303   0   0
#>   0.89954931  0.82804864  0.865489043   0   0
#>   0.87398760  0.73412140  0.798009024   0   0
#>   0.85835829  0.83903425  0.772886862   0   0
#>   0.73606823  0.50957003  0.804829616   0   0
#>   0.71312160  0.47475958  0.728586398   0   0
#>   0.81452459  0.56621278  0.745892376   0   0
#>   0.72858640  0.47531326  0.795270254   0   0
#>   0.34413001  0.28795415  0.449575215   0   0
#>   0.84856945  0.57256400  0.830200303   0   0
#>   0.83338738  0.64525019  0.833036756   0   0
#>   0.88511432  0.64540439  0.833387384   0   0
#>   0.60044041  0.43155542  0.604400619   0   0
#>   0.71701174  0.36986314  0.739921040   0   0
#>   0.45732027  0.28795415  0.449575215   0   0
#>   0.94241098  0.69551117  0.885114321   0   0
#>   0.92351259  0.79842531  0.899549309   0   0
#>   1.30653302  0.95332691  0.901003195   0   0
#>   0.22118233  0.22645349  0.235095254   0   0
#>   0.33087070  0.23032091  0.313586627   0   0
#>   0.59242831  0.25954814  0.468178943   0   0
#>   0.90221823  0.55540049  0.848569449   0   0
#>   0.44957521  0.27205090  0.455192693   0   0
#>   1.03398475  0.79832541  0.923512592   0   0
#>   0.22125891  0.33268154  0.330870705   0   0
#>   0.81049788  0.47531326  0.795270254   0   0
#>   0.92639209  0.64540439  0.833387384   0   0
#>  -0.34375019  0.05391676 -2.479519160   0   1
#>   0.17105752  0.22129109  0.153049564   0   0
#>   0.60362551  0.36573371  0.600440414   0   0
#>   0.71247338  0.30392088  0.723243339   0   0
#>   0.82294037  0.45715870  0.810497883   0   0
#>   0.89585071  0.39825519  0.923384910   0   0
#>   0.46817894  0.25292216  0.469565215   0   0
#>   0.45519269  0.23032091  0.313586627   0   0
#>   0.73992104  0.37851444  0.807439531   0   0
#>   0.92636012  0.55738551  0.915699210   0   0
#>   0.89656325  0.45715870  0.810497883   0   0
#>   0.57499666  0.16589077  0.604933120   0   0
#>   0.80743953  0.26412454  0.847371015   0   0
#>   0.91569921  0.54959891  0.902218226   0   0
#>  -0.14356435  0.22129109  0.153049564   0   0
#>   0.15304956  0.38557553  0.129599348   0   0
#>   0.34816924  0.22704184  0.311241913   0   0
#>   0.61140976  0.32335528  0.592428305   0   0
#>   0.82953708  0.39825519  0.923384910   0   0
#>   1.14081867  0.55738551  0.915699210   0   0
#>   0.70052992  0.16589077  0.604933120   0   0
#>   0.72324334  0.26412454  0.847371015   0   0
#>   0.31124191  0.23032091  0.313586627   0   0
#>   0.56013718  0.10392051  0.512817872   0   0
#>   0.70549268  0.16308987  0.752242151   0   0
#>   0.20401390  0.07037983  0.209017591   0   0
#>   0.37948574  0.07256636  0.361932429   0   0
#>   0.68451368  0.07116077  0.682782378   0   0
#>   0.17659535  0.22645349  0.235095254   0   0
#>   0.41836460  0.10392051  0.512817872   0   0
#>   0.36583897  0.07256636  0.361932429   0   0
#>   0.51554862  0.05731579  0.365838970   0   0
#>  -0.13448137  0.19131889  0.171057520   0   0
#>   0.20231684  0.07037983  0.209017591   0   0
#>   0.18952175  0.05731579  0.365838970   0   0
#>   0.68278238 -0.55295363  0.920752389   0   0
#>   0.75224215  0.17841791  0.861526475   0   0
#>   0.88115829  0.26412454  0.847371015   0   0
#>   0.53391922  0.10392051  0.512817872   0   0
#>   1.04700126  0.39825519  0.923384910   0   0
#>   0.86811281  0.17841791  0.861526475   0   0
#>   0.46663446  0.94496062  0.481455777   0   0
#>   0.24205931  0.77757850  0.254397729   0   0
#>   0.43033176  0.94496062  0.481455777   0   0
#>   0.32480780  0.77757850  0.254397729   0   0
#>   0.46753990  0.81960002  0.478827887   0   0
#>   0.58414119  0.90916031  0.588781368   0   0
#>   0.45555437  0.64896158  0.533084168   0   0
#>   0.52455675  0.74721614  0.567155301   0   0
#>   0.38236474  0.66797921  0.447118523   0   0
#>   0.44612194  0.69954271  0.455554373   0   0
#>   0.66294833  0.84576406  0.669184457   0   0
#>   0.58878137  0.88569412  0.573602877   0   0
#>   0.58205058  0.74721614  0.567155301   0   0
#>   0.14621907  6.70442092 -2.085525390   0   1
#>   0.54872755  0.64896158  0.533084168   0   0
#>   0.61628879  0.70054478  0.657023513   0   0
#>   0.23371274  0.78274277  0.199422937   0   0
#>   0.17523092  0.83119617  0.146219069   0   0
#>   0.36217640  0.57608080  0.379850301   0   0
#>   0.69887180  0.84149910  0.770100039   0   0
#>   0.19942294  0.79880749  0.175230924   0   0
#>   0.68888923  0.84576406  0.669184457   0   0
#>   0.69812376  0.85727512  0.688889233   0   0
#>   0.63213993  0.95782506  0.698123761   0   0
#>  -1.09280467 42.76174565 -1.194965150   0   1
#>   0.72374954  0.74698157  0.767480409   0   0
#>   0.75483403  0.84149910  0.770100039   0   0
#>   0.74692449  0.94102692  0.759985780   0   0
#>   0.04010243  0.68526012  0.046899902   0   0
#>   0.24486502  0.58874362  0.362176402   0   0
#>   0.51196261  0.62761589  0.576475588   0   0
#>   0.60412896  0.65738781  0.616288792   0   0
#>   0.51331383  0.51029175  0.505807011   0   0
#>  -0.28327991  0.74896123 -0.001822419   0   0
#>   0.63523469  0.58191994  0.651499863   0   0
#>   0.82239020  1.00836645  0.772181882   0   0
#>   0.77218188 14.95752273  2.773198191   0   1
#>   0.76748041  0.73614463  0.795929969   0   0
#>   0.70869756  0.73300538  0.723749536   0   0
#>   0.86548904  0.84185166  0.858358290   0   0
#>   0.71454178  0.71119350  0.708697562   0   0
#>   0.15815329  0.58402992  0.020283547   0   0
#>   0.42953521  0.44934555  0.377349833   0   0
#>   0.40652531  0.51029175  0.505807011   0   0
#>   0.70475889  0.57704586  0.668133719   0   0
#>   0.66813372  0.58191994  0.651499863   0   0
#>   0.77034247  0.63619911  0.774124644   0   0
#>   0.87827410  0.95332691  0.901003195   0   0
#>   0.37541272  0.50361368  0.406525314   0   0
#>   0.53040423  0.50220593  0.513313832   0   0
#>   0.79682216  0.73412140  0.798009024   0   0
#>   0.49427929  0.41750685  0.429535210   0   0
#>   0.60440062  0.45558547  0.654668442   0   0
#>   0.77412464  0.57531209  0.706711345   0   0
#>   0.37734983  0.45494046  0.375412719   0   0
#>   0.74589238  0.56463215  0.736068228   0   0
#>   0.81406177  0.64525019  0.833036756   0   0
#>   0.49611029  0.36573371  0.600440414   0   0
#>   0.87398760  0.79842531  0.899549309   0   0
#>   0.85835829  0.91472030  0.878274097   0   0
#>   0.73606823  0.56861567  0.713121600   0   0
#>   0.81452459  0.62370739  0.814061769   0   0
#>   0.72858640  0.46006476  0.717011737   0   0
#>   0.65466844  0.46006476  0.717011737   0   0
#>   0.34413001  0.33268154  0.330870705   0   0
#>   0.36739832  0.34411433  0.457320266   0   0
#>   0.47016950  0.38082984  0.496110294   0   0
#>   0.88511432  0.71341641  0.873987600   0   0
#>   0.94241098  0.71862416  1.033984753   0   0
#>   0.92351259  1.30886742  1.306533020   0   0
#>   1.30653302  4.16719743  4.230115391   0   1
#>   0.59242831  0.35859692  0.603625512   0   0
#>   0.80482962  0.55540049  0.848569449   0   0
#>   0.79527025  0.50957003  0.804829616   0   0
#>   1.03398475  1.28199366  5.883108996   0   1
#>   0.10437510  0.38557553  0.129599348   0   0
#>   0.22125891  0.38412538  0.221182330   0   0
#>   0.23509525  0.22704184  0.311241913   0   0
#>   0.92639209  0.66585459  0.942410982   0   0
#>   0.17105752  0.18958354  0.176595348   0   0
#>   0.60362551  0.34272479  0.712473376   0   0
#>   0.82294037  0.49463874  0.895850712   0   0
#>   0.46817894  0.27205090  0.455192693   0   0
#>   0.73992104  0.34272479  0.712473376   0   0
#>   0.92636012  0.58050396  0.926392091   0   0
#>   0.89656325  0.54959891  0.902218226   0   0
#>   0.57499666  0.18078244  0.560137178   0   0
#>  -0.14356435  0.39661167 -0.343750186   0   0
#>   0.34816924  0.12411393  0.379485740   0   0
#>   0.46956521  0.18078244  0.560137178   0   0
#>   0.61140976  0.22843402  0.700529919   0   0
#>   0.82953708  0.37851444  0.807439531   0   0
#>   1.14081867  0.27113183  3.758446803   0   1
#>   0.72324334  0.22789726  0.705492682   0   0
#>   0.60493312  0.07937495  0.684513684   0   0
#>   0.70549268  0.22843402  0.700529919   0   0
#>   0.20401390  0.18958354  0.176595348   0   0
#>   0.41836460  0.12411393  0.379485740   0   0
#>   0.20901759  0.07256636  0.361932429   0   0
#>   0.51554862  0.05010565  0.533919222   0   0
#>  -0.13448137  0.05589100  0.202316841   0   0
#>   0.20231684 -1.08097720  0.189521750   0   0
#>   0.18952175 -4.25250044 -0.054024509   0   1
#>   0.68278238  0.05010565  0.533919222   0   0
#>   0.75224215  0.07937495  0.684513684   0   0
#>   0.86152647  0.22323511  0.881158288   0   0
#>   0.88115829  0.33348733  1.047001264   0   0
#>   1.04700126  0.34922860  1.140818667   0   0
#>   0.92075239  0.16401169  0.868112812   0   0
#>   0.86811281 -4.13785657 12.100276354   0   1
#>   0.46663446  0.82790393  0.475735212   0   0
#>   0.24359536  7.39305599 -2.092315600   0   1
#>   0.24205931  0.76938209  0.233712742   0   0
#>   0.46753990  0.73015783  0.476802933   0   0
#>   0.57564362  0.77942275  0.582050579   0   0
#>   0.52455675  0.64894261  0.548727549   0   0
#>   0.44711852  0.68990751  0.446121944   0   0
#>   0.04010243  0.69430677  0.228791140   0   0
#>  -0.28327991  2.00409659 -5.270854500   0   1
#>   0.63523469  0.49510667  0.530404234   0   0
#>   0.71454178  0.65075592  0.770342473   0   0
#>   0.15815329  0.57877673  0.230097671   0   0
#>   0.27054685  0.38416485  0.221258911   0   0
#>   0.81452459  0.57256400  0.830200303   0   0
#>   0.34413001  0.38326045  0.367398318   0   0
#>   0.47016950  0.34411433  0.457320266   0   0
#>   0.10437510  0.38358327 -0.143564352   0   0
#>   0.82294037  0.40332778  0.829537078   0   0
#>   0.89585071  0.52363853  0.896563255   0   0
#>   0.92636012  0.33947281 15.827923764   0   1
#>   0.57499666  0.25931030  0.611409765   0   0
#>   0.34816924  0.10749240  0.204013898   0   0
#>   0.46956521  0.13842404  0.418364596   0   0
#>   0.51554862 -5.84876719  0.531511517   0   1
#>  -0.13448137 -0.90889633 -4.795389702   0   1
#>   0.92075239 -3.20587890  1.550312878   0   1
#> 
#> $x
#>                [,1]       [,2]
#>   [1,] 0.9320831406 0.29343386
#>   [2,] 0.6858758383 0.00000000
#>   [3,] 0.0728951147 0.79090864
#>   [4,] 0.9013516232 0.42975245
#>   [5,] 0.5256790297 0.33070852
#>   [6,] 0.3478301445 0.40523927
#>   [7,] 0.6102308867 0.12076998
#>   [8,] 0.3447138261 0.86658991
#>   [9,] 0.7960683717 0.04017765
#>  [10,] 0.7237498023 0.60603088
#>  [11,] 0.8376727524 0.59882759
#>  [12,] 0.0643051965 0.45172619
#>  [13,] 0.6650703676 0.49284798
#>  [14,] 0.9327458737 0.19276369
#>  [15,] 0.2491474388 0.55183000
#>  [16,] 0.6388208140 0.41177121
#>  [17,] 0.7360947736 0.42029672
#>  [18,] 0.0004654966 0.60095039
#>  [19,] 0.6609343789 0.40184143
#>  [20,] 0.1735863674 0.49164631
#>  [21,] 0.3950913064 0.31692778
#>  [22,] 0.8366833721 0.55035885
#>  [23,] 0.7429031572 0.92431579
#>  [24,] 0.9099384676 0.72737438
#>  [25,] 0.5024467113 0.67934023
#>  [26,] 0.8762108796 0.04398199
#>  [27,] 0.3035614852 0.27393735
#>  [28,] 0.2841240813 0.10317732
#>  [29,] 1.0000000000 0.73513725
#>  [30,] 0.9910335084 0.53097616
#>  [31,] 0.1627892487 0.98753316
#>  [32,] 0.9832767746 0.33088983
#>  [33,] 0.2864820551 0.37508600
#>  [34,] 0.7992386080 0.73059644
#>  [35,] 0.5083524714 0.09139570
#>  [36,] 0.1163884527 0.58455432
#>  [37,] 0.4333691279 0.76703888
#>  [38,] 0.4211508341 0.46857894
#>  [39,] 0.7306851647 0.05858514
#>  [40,] 0.0000000000 0.43018464
#>  [41,] 0.6464687277 0.73825932
#>  [42,] 0.6830194157 0.76773958
#>  [43,] 0.9257942391 0.80238815
#>  [44,] 0.7616426505 0.52409888
#>  [45,] 0.5444747351 0.99892072
#>  [46,] 0.5067127083 0.64085980
#>  [47,] 0.5526730694 0.57536616
#>  [48,] 0.6361404762 0.86905576
#>  [49,] 0.4754280624 0.93507666
#>  [50,] 0.5373978378 0.58996027
#>  [51,] 0.5167828198 0.76634629
#>  [52,] 0.3152458324 0.78217429
#>  [53,] 0.4738360077 0.18617326
#>  [54,] 0.0240708916 0.11672718
#>  [55,] 0.3797033383 0.41911477
#>  [56,] 0.6767883619 0.09303242
#>  [57,] 0.2061666000 0.64019830
#>  [58,] 0.8150407644 0.97636019
#>  [59,] 0.1499548879 0.27512420
#>  [60,] 0.2294124287 0.07324186
#>  [61,] 0.2604417267 0.76460365
#>  [62,] 0.5455459787 0.78937899
#>  [63,] 0.4958147194 0.84148725
#>  [64,] 0.5695686466 0.44994199
#>  [65,] 0.6151227780 0.85438921
#>  [66,] 0.6187003792 0.50205670
#>  [67,] 0.0986972045 0.12165868
#>  [68,] 0.7945616719 0.81022803
#>  [69,] 0.4127326362 0.68578608
#>  [70,] 0.0851480491 0.95779693
#>  [71,] 0.6115644243 1.00000000
#>  [72,] 0.3335945662 0.50846293
#>  [73,] 0.6626284435 0.80045479
#>  [74,] 0.8037930344 0.72622822
#>  [75,] 0.0221526268 0.28716866
#>  [76,] 0.8537390501 0.92477783
#>  [77,] 0.6592149878 0.36110097
#>  [78,] 0.8890089795 0.51946094
#>  [79,] 0.2955681278 0.18335051
#>  [80,] 0.4440672753 0.45838370
#>  [81,] 0.4226263116 0.52134545
#>  [82,] 0.9899235280 0.80538069
#>  [83,] 0.5931613656 0.78579195
#>  [84,] 0.1820719555 0.65596198
#>  [85,] 0.4576064067 0.87269189
#>  [86,] 0.7857303443 0.41354876
#>  [87,] 0.2164850756 0.39571346
#>  [88,] 0.4773776941 0.85572863
#>  [89,] 0.6792979125 0.50213016
#>  [90,] 0.8858693880 0.80721824
#>  [91,] 0.2832531542 0.66690546
#>  [92,] 0.7018232809 0.56594370
#>  [93,] 0.7563972900 0.66140742
#>  [94,] 0.5994768346 0.78070193
#>  [95,] 0.6509760081 0.68313210
#>  [96,] 0.9747541543 0.78986000
#>  [97,] 0.1208218221 0.28575779
#>  [98,] 0.6813494192 0.82526283
#>  [99,] 0.9160673498 0.66064019
#> [100,] 0.7657701867 0.84826841
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,32,4), [7,2,5]
#> 2: (86,1,4), [1,3,15]
#> 3: (78,86,4), [2,4,10]
#> 4: (4,30,78), [14,3,7]
#> 5: (14,32,1), [1,6,0]
#> 6: (77,14,1), [5,15,34]
#> 7: (4,32,30), [44,4,1]
#> 8: (86,17,77), [16,15,12]
#> 9: (44,86,22), [10,13,12]
#> 10: (86,78,22), [11,9,3]
#> 11: (78,11,22), [13,10,25]
#> 12: (44,17,86), [8,9,17]
#> 13: (11,44,22), [9,11,27]
#> 14: (78,30,99), [43,25,4]
#> 15: (77,1,86), [2,8,6]
#> 16: (17,19,77), [20,8,22]
#> 17: (17,44,89), [19,18,12]
#> 18: (13,17,89), [17,26,22]
#> 19: (92,89,44), [17,23,30]
#> 20: (16,77,19), [16,21,37]
#> 21: (19,13,16), [29,20,22]
#> 22: (13,19,17), [16,18,21]
#> 23: (44,10,92), [31,19,27]
#> 24: (74,93,11), [33,46,39]
#> 25: (78,99,11), [46,11,14]
#> 26: (89,66,13), [29,18,30]
#> 27: (44,11,10), [33,23,13]
#> 28: (14,9,26), [58,0,35]
#> 29: (66,16,13), [21,26,38]
#> 30: (89,92,66), [32,26,19]
#> 31: (95,92,10), [23,48,54]
#> 32: (47,66,92), [30,54,53]
#> 33: (93,10,11), [27,24,48]
#> 34: (14,77,56), [51,40,6]
#> 35: (14,39,9), [59,28,40]
#> 36: (5,16,64), [38,76,37]
#> 37: (5,77,16), [20,36,52]
#> 38: (64,16,66), [29,53,36]
#> 39: (34,93,74), [24,64,45]
#> 40: (14,56,39), [60,35,34]
#> 41: (99,24,74), [47,46,42]
#> 42: (99,29,24), [49,41,43]
#> 43: (99,30,29), [44,42,14]
#> 44: (29,30,32), [7,0,43]
#> 45: (42,93,34), [39,68,69]
#> 46: (74,11,99), [25,41,24]
#> 47: (74,24,90), [57,64,41]
#> 48: (95,10,93), [33,69,31]
#> 49: (24,29,96), [67,56,42]
#> 50: (56,7,2), [77,60,51]
#> 51: (7,56,77), [34,52,50]
#> 52: (5,7,77), [51,37,74]
#> 53: (47,64,66), [38,32,62]
#> 54: (95,47,92), [32,31,63]
#> 55: (47,81,80), [89,62,86]
#> 56: (24,96,43), [66,57,49]
#> 57: (90,24,43), [56,84,47]
#> 58: (2,26,9), [28,59,0]
#> 59: (2,9,39), [35,60,58]
#> 60: (56,2,39), [59,40,50]
#> 61: (68,34,90), [64,101,68]
#> 62: (80,64,47), [53,55,76]
#> 63: (50,47,95), [54,71,86]
#> 64: (74,90,34), [61,39,47]
#> 65: (76,43,82), [66,119,84]
#> 66: (82,43,96), [56,67,65]
#> 67: (82,96,29), [49,0,66]
#> 68: (42,34,68), [61,83,45]
#> 69: (42,95,93), [48,45,72]
#> 70: (76,100,68), [83,101,99]
#> 71: (95,46,50), [78,63,80]
#> 72: (42,41,95), [79,69,81]
#> 73: (7,53,35), [128,77,74]
#> 74: (53,7,5), [52,94,73]
#> 75: (80,38,55), [109,95,89]
#> 76: (80,5,64), [36,62,85]
#> 77: (7,35,2), [133,50,73]
#> 78: (81,50,46), [71,90,86]
#> 79: (95,41,25), [91,80,72]
#> 80: (95,25,46), [106,71,79]
#> 81: (41,42,73), [88,92,72]
#> 82: (98,42,100), [83,100,88]
#> 83: (42,68,100), [70,82,68]
#> 84: (76,90,43), [57,65,101]
#> 85: (80,21,5), [94,76,95]
#> 86: (47,50,81), [78,55,63]
#> 87: (65,73,98), [88,112,97]
#> 88: (98,73,42), [81,82,87]
#> 89: (80,81,38), [98,75,55]
#> 90: (69,81,46), [78,106,114]
#> 91: (41,94,25), [103,79,92]
#> 92: (41,73,94), [97,91,81]
#> 93: (28,53,79), [120,150,128]
#> 94: (21,53,5), [74,85,129]
#> 95: (80,55,21), [108,85,75]
#> 96: (94,83,62), [111,102,104]
#> 97: (65,94,73), [92,87,104]
#> 98: (72,38,81), [89,114,109]
#> 99: (23,100,76), [70,118,100]
#> 100: (23,98,100), [82,99,113]
#> 101: (76,68,90), [61,84,70]
#> 102: (94,62,51), [123,103,96]
#> 103: (94,51,25), [105,91,102]
#> 104: (65,83,94), [96,97,111]
#> 105: (25,51,37), [124,115,103]
#> 106: (69,46,25), [80,115,90]
#> 107: (21,6,33), [126,121,108]
#> 108: (6,21,55), [95,116,107]
#> 109: (72,55,38), [75,98,116]
#> 110: (63,62,65), [111,125,123]
#> 111: (83,65,62), [110,96,104]
#> 112: (48,65,98), [87,113,132]
#> 113: (23,48,98), [112,100,117]
#> 114: (72,81,69), [90,135,98]
#> 115: (25,37,69), [142,106,105]
#> 116: (55,72,6), [126,108,109]
#> 117: (71,48,23), [113,127,132]
#> 118: (58,23,76), [99,119,127]
#> 119: (58,76,82), [65,0,118]
#> 120: (53,27,79), [130,93,129]
#> 121: (21,33,27), [141,129,107]
#> 122: (91,15,72), [139,135,153]
#> 123: (63,51,62), [102,110,124]
#> 124: (63,37,51), [105,123,131]
#> 125: (49,63,65), [110,148,145]
#> 126: (6,72,33), [140,107,116]
#> 127: (71,23,58), [118,0,117]
#> 128: (28,35,53), [73,93,149]
#> 129: (53,21,27), [121,120,94]
#> 130: (59,79,27), [120,159,167]
#> 131: (88,37,63), [124,145,137]
#> 132: (71,65,48), [112,117,143]
#> 133: (35,60,2), [0,77,149]
#> 134: (67,60,79), [150,167,173]
#> 135: (91,72,69), [114,136,122]
#> 136: (69,52,91), [158,135,142]
#> 137: (85,37,88), [131,138,155]
#> 138: (88,49,85), [144,137,145]
#> 139: (72,15,87), [152,140,122]
#> 140: (72,87,33), [141,126,139]
#> 141: (87,27,33), [121,140,159]
#> 142: (69,37,52), [147,136,115]
#> 143: (71,45,65), [148,132,0]
#> 144: (8,85,49), [138,181,155]
#> 145: (49,88,63), [131,125,138]
#> 146: (15,57,36), [160,161,153]
#> 147: (37,8,52), [154,142,155]
#> 148: (45,49,65), [125,143,156]
#> 149: (35,28,60), [150,133,128]
#> 150: (60,28,79), [93,134,149]
#> 151: (97,59,87), [159,164,163]
#> 152: (20,87,15), [139,161,169]
#> 153: (57,15,91), [122,157,146]
#> 154: (61,52,8), [147,179,158]
#> 155: (37,85,8), [144,147,137]
#> 156: (31,49,45), [148,0,181]
#> 157: (91,84,57), [160,153,162]
#> 158: (91,52,61), [154,162,136]
#> 159: (59,27,87), [141,151,130]
#> 160: (36,57,84), [157,166,146]
#> 161: (15,36,20), [165,152,146]
#> 162: (91,61,84), [177,157,158]
#> 163: (59,97,67), [170,167,151]
#> 164: (87,12,97), [168,151,169]
#> 165: (12,20,36), [161,180,169]
#> 166: (84,3,36), [176,160,177]
#> 167: (67,79,59), [130,163,134]
#> 168: (75,97,12), [164,171,170]
#> 169: (87,20,12), [165,164,152]
#> 170: (75,67,97), [163,168,174]
#> 171: (40,75,12), [168,172,175]
#> 172: (18,40,12), [171,180,0]
#> 173: (54,60,67), [134,174,0]
#> 174: (54,67,75), [170,175,173]
#> 175: (54,75,40), [171,0,174]
#> 176: (36,3,18), [182,180,166]
#> 177: (84,61,3), [178,166,162]
#> 178: (31,3,61), [177,179,183]
#> 179: (31,61,8), [154,181,178]
#> 180: (18,12,36), [165,176,172]
#> 181: (31,8,49), [144,156,179]
#> 182: (70,18,3), [176,183,0]
#> 183: (70,3,31), [178,0,182]
#> boundary nodes:  14 32 29 82 58 71 45 31 70 18 40 54 60 2 26 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.128396059 0.139739705 0.116751187 0.102672317 0.063432942 0.100672351 
#>                                                                         
#> 0.200236623 0.097029055 0.145990453 0.060767180 0.048478838 0.050092163 
#>                                                                         
#> 0.079502806 0.149775456 0.189332329 0.040776725 0.085224882 0.099612227 
#>                                                                         
#> 0.085220197 0.077393039 0.045691948 0.102576487 0.099832413 0.047275536 
#>                                                                         
#> 0.114150453 0.080232753 0.104837809 0.268643950 0.067924955 0.079075100 
#>                                                                         
#> 0.054620465 0.092500280 0.080300016 0.063964679 0.106151621 0.090395474 
#>                                                                         
#> 0.204357891 0.081378733 0.083392829 0.064283899 0.060265518 0.142504550 
#>                                                                         
#> 0.245272151 0.071623110 0.149447543 0.066512382 0.050537390 0.076671193 
#>                                                                         
#> 0.073756998 0.115653766 0.126557029 0.145866263 0.064199073 0.079768822 
#>                                                                         
#> 0.107636464 0.047707648 0.059433484 0.055311177 0.100867098 0.226325446 
#>                                                                         
#> 0.064533705 0.127054788 0.199669026 0.155607844 0.038716174 0.038549661 
#>                                                                         
#> 0.115466798 0.040215961 0.131312841 0.133717977 0.031079154 0.052787136 
#>                                                                         
#> 0.146130771 0.140317639 0.066168731 0.178290227 0.153551754 0.103339102 
#>                                                                         
#> 0.047750308 0.116585643 0.087499272 0.091357309 0.036848672 0.088179243 
#>                                                                         
#> 0.008111317 0.083416567 0.068357938 0.095222161 0.072348561 0.062941793 
#>                                                                         
#> 0.164737990 0.083832467 0.104200653 0.120216226 0.110836856 0.181112913 
#>                                                                         
#> 0.090938818 0.102580487 0.094931659 0.120003640 0.141453154 0.088951977 
#>                                                                         
#> 0.101123120 0.097170011 0.025629108 0.462302081 0.128454243 0.194181186 
#>                                                                         
#> 0.119626234 0.099008213 0.064880431 0.159496721 0.072973072 0.119089015 
#>                                                                         
#> 0.160874183 0.023296803 0.105623714 0.089411184 0.161322896 0.062365828 
#>                                                                         
#> 0.080985849 0.137724494 0.120023168 0.113057370 0.094039927 0.028793150 
#>                                                                         
#> 0.057551867 0.149705451 0.109103156 0.133958327 0.165583864 0.175327230 
#>                                                                         
#> 0.210888067 0.172120958 0.116344277 0.169870590 0.067817375 0.139393897 
#>                                                                         
#> 0.182353026 0.144721488 0.203298271 0.189382375 0.132298845 0.142674377 
#>                                                                         
#> 0.147569121 0.216199348 0.090553595 0.281133184 0.135237123 0.136955854 
#>                                                                         
#> 0.113143914 0.143748870 0.016987698 0.073002374 0.024240673 0.101529071 
#>                                                                         
#> 0.131828506 0.094522347 0.090270268 0.159160921 0.060597578 0.106021807 
#>                                                                         
#> 0.274700880 0.204849305 0.139184132 0.006310634 0.242553936 0.130028116 
#>                                                                         
#> 0.067015028 0.112225222 0.404593177 0.122010303 0.115307900 0.090030699 
#>                                                                         
#> 0.093475202 0.136950915 0.098660439 0.127744142 0.159624916 0.117288696 
#>                                                                         
#> 0.146904698 0.021702575 0.070962505 0.119360560 0.129193039 0.128932607 
#>                                                                         
#> 0.090472128 0.106028580 0.075397383 0.125784962 0.148577691 0.150329129 
#>                                                                         
#> 0.064260502 0.142025290 0.151530293 0.140820439 0.057547481 0.025081989 
#>                                                                         
#> 0.104166331 0.063321505 0.149694431 0.054624432 0.071872878 0.089958940 
#>                                                                         
#> 0.079410992 0.121871300 0.083930832 0.075330014 0.111636759 0.089945340 
#>                                                                         
#> 0.123200438 0.034762479 0.096212436 0.116620546 0.151584836 0.064484866 
#>                                                                         
#> 0.244577126 0.166247681 0.072031023 0.078011693 0.204844969 0.207072310 
#>                                                                         
#> 0.191561867 0.153611182 0.133230539 0.206310549 0.130848822 0.026051024 
#>                                                                         
#> 0.162564069 0.137092540 0.145654257 0.095784098 0.136732693 0.279530158 
#>                                                                         
#> 0.145745351 0.096599957 0.081581916 0.133304803 0.381855322 0.100325961 
#>                                                                         
#> 0.097022560 0.101771286 0.161799302 0.162160850 0.098679282 0.162306406 
#>                                                                         
#> 0.074789079 0.170452277 0.314380324 0.117076739 0.173580480 0.243379483 
#>                                                                         
#> 0.218457824 0.317009010 0.167337491 0.083140840 0.147931965 0.147078902 
#>                                                                         
#> 0.321179991 0.106899873 0.106606554 0.067672453 0.091100485 0.072105962 
#>                                                                         
#> 0.195350566 0.021126309 0.046957837 0.151265170 0.152635097 0.072026994 
#>                                                                         
#> 0.100162550 0.100544086 0.224537697 0.108397436 0.079371970 0.067098370 
#>                                                                         
#> 0.098266522 0.031013043 0.105087637 0.170766380 0.209895498 0.366756853 
#> 
#> $alpha
#> [1] 0.2522949
#> 
#> attr(,"class")
#> [1] "scree"