Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    34   99 0.319364664 0.92745331 0.318085050 0.83410145  0.312224114
#>    34    1 0.319364664 0.92745331 0.344803363 0.84680901  0.360260343
#>    65   50 0.350738059 0.73668302 0.405913711 0.87201231  0.402058411
#>    17   34 0.202246247 0.88123708 0.319364664 0.92745331  0.252201094
#>    50   75 0.405913711 0.87201231 0.463449311 0.92122834  0.396708052
#>    17   14 0.202246247 0.88123708 0.269130021 0.99115797  0.236368248
#>    99   65 0.318085050 0.83410145 0.350738059 0.73668302  0.354439971
#>    59   65 0.163880024 0.81531479 0.350738059 0.73668302  0.249249937
#>    34   50 0.319364664 0.92745331 0.405913711 0.87201231  0.391447485
#>    50   44 0.405913711 0.87201231 0.440802661 0.66171352  0.480280805
#>    17   59 0.202246247 0.88123708 0.163880024 0.81531479  0.242312264
#>    17   67 0.202246247 0.88123708 0.164802363 0.98364114  0.138135092
#>    14   34 0.269130021 0.99115797 0.319364664 0.92745331  0.359300951
#>    34   89 0.319364664 0.92745331 0.464730460 0.95580398  0.390835629
#>    89   75 0.464730460 0.95580398 0.463449311 0.92122834  0.463294906
#>     2   79 0.550956198 0.89005271 0.478250328 0.92981364  0.497758310
#>    21   45 0.449852970 0.99301774 0.240357038 1.00000000  0.356062677
#>    21   89 0.449852970 0.99301774 0.464730460 0.95580398  0.576217341
#>    45   67 0.240357038 1.00000000 0.164802363 0.98364114  0.213095652
#>    40   44 0.331518855 0.63010726 0.440802661 0.66171352  0.377201054
#>    59    7 0.163880024 0.81531479 0.080060133 0.73502438  0.178455172
#>    43   41 0.570689686 0.87103967 0.665094146 0.92758251  0.602945829
#>    87   43 0.590170156 0.82743770 0.570689686 0.87103967  0.493039789
#>    10   60 0.073201781 0.88280553 0.016906306 0.81584939  0.090377300
#>     6   35 0.087447929 0.64344374 0.158926877 0.56442975  0.154825071
#>    50   87 0.405913711 0.87201231 0.590170156 0.82743770  0.487664495
#>    60    7 0.016906306 0.81584939 0.080060133 0.73502438  0.090366178
#>    40   59 0.331518855 0.63010726 0.163880024 0.81531479  0.219582314
#>     6   92 0.087447929 0.64344374 0.093664681 0.66284925  0.001481859
#>    87   44 0.590170156 0.82743770 0.440802661 0.66171352  0.481861060
#>    39   63 0.511248648 0.58927462 0.565504476 0.68829859  0.507237933
#>    35   40 0.158926877 0.56442975 0.331518855 0.63010726  0.216285579
#>    60   36 0.016906306 0.81584939 0.000000000 0.77575771  0.051381121
#>    92    7 0.093664681 0.66284925 0.080060133 0.73502438  0.003292476
#>    35   84 0.158926877 0.56442975 0.196093551 0.51396227  0.247661889
#>    73   39 0.358015069 0.60447089 0.511248648 0.58927462  0.433354996
#>     9   73 0.316779684 0.47303579 0.358015069 0.60447089  0.284758423
#>    89    2 0.464730460 0.95580398 0.550956198 0.89005271  0.574599943
#>    87   71 0.590170156 0.82743770 0.620659969 0.68149268  0.609579683
#>    63   71 0.565504476 0.68829859 0.620659969 0.68149268  0.601567567
#>    63   93 0.565504476 0.68829859 0.544106450 0.58969567  0.550611521
#>    39   93 0.511248648 0.58927462 0.544106450 0.58969567  0.526965666
#>    84    9 0.196093551 0.51396227 0.316779684 0.47303579  0.277610805
#>     9   96 0.316779684 0.47303579 0.388504122 0.45311624  0.370752967
#>    11   84 0.126976444 0.43778443 0.196093551 0.51396227  0.130169267
#>    87   23 0.590170156 0.82743770 0.675419183 0.72011230  0.674818944
#>    90   71 0.750211405 0.68281735 0.620659969 0.68149268  0.685922505
#>    96   39 0.388504122 0.45311624 0.511248648 0.58927462  0.428993807
#>    23   71 0.675419183 0.72011230 0.620659969 0.68149268  0.685792363
#>    93   70 0.544106450 0.58969567 0.528794122 0.50917568  0.599863692
#>    70   39 0.528794122 0.50917568 0.511248648 0.58927462  0.438574038
#>    80   78 0.749206139 0.85235847 0.795554876 0.87804735  0.745215892
#>     5   87 0.684431950 0.90207472 0.590170156 0.82743770  0.672822659
#>    11    9 0.126976444 0.43778443 0.316779684 0.47303579  0.229536544
#>    66   11 0.064687161 0.51903636 0.126976444 0.43778443  0.129774448
#>     6   66 0.087447929 0.64344374 0.064687161 0.51903636  0.094392143
#>     6   36 0.087447929 0.64344374 0.000000000 0.77575771 -0.101374112
#>    80   23 0.749206139 0.85235847 0.675419183 0.72011230  0.703644843
#>    83    3 0.774688881 0.82321841 0.761596075 0.84080656  0.708493819
#>    83   23 0.774688881 0.82321841 0.675419183 0.72011230  0.740435879
#>    70   27 0.528794122 0.50917568 0.474154623 0.43693288  0.516773820
#>    70   96 0.528794122 0.50917568 0.388504122 0.45311624  0.444734035
#>    11   95 0.126976444 0.43778443 0.141701822 0.34328150  0.231126433
#>    77   96 0.373287922 0.36956137 0.388504122 0.45311624  0.340325194
#>    71   30 0.620659969 0.68149268 0.600933809 0.64549187  0.673350530
#>    78   41 0.795554876 0.87804735 0.665094146 0.92758251  0.759187800
#>    80    3 0.749206139 0.85235847 0.761596075 0.84080656  0.772522759
#>    18   78 0.832330316 0.80257071 0.795554876 0.87804735  0.813200134
#>    30   93 0.600933809 0.64549187 0.544106450 0.58969567  0.610842699
#>    82    4 0.619497870 0.41462239 0.664216228 0.44892703  0.599839888
#>    29   33 0.455503619 0.33626992 0.551179583 0.42633543  0.505436140
#>    27   96 0.474154623 0.43693288 0.388504122 0.45311624  0.423498649
#>    90   18 0.750211405 0.68281735 0.832330316 0.80257071  0.780908482
#>    19   90 0.900917128 0.72823690 0.750211405 0.68281735  0.824834448
#>    26   90 0.902980245 0.71761406 0.750211405 0.68281735  0.837691719
#>    51   70 0.674265294 0.55499947 0.528794122 0.50917568  0.609228008
#>    33   82 0.551179583 0.42633543 0.619497870 0.41462239  0.596041182
#>    77   29 0.373287922 0.36956137 0.455503619 0.33626992  0.430603859
#>    29   38 0.455503619 0.33626992 0.457780416 0.30395176  0.543872609
#>    77   95 0.373287922 0.36956137 0.141701822 0.34328150  0.257303139
#>    16   76 0.033782071 0.32910251 0.110336345 0.33461694  0.067551113
#>    13   41 0.886276338 0.94994035 0.665094146 0.92758251  0.769017785
#>    78   94 0.795554876 0.87804735 0.882722195 0.86833393  0.843342859
#>    88   94 0.932702891 0.81776290 0.882722195 0.86833393  0.881465964
#>    77   42 0.373287922 0.36956137 0.379801397 0.26304469  0.400154910
#>    76   54 0.110336345 0.33461694 0.143674555 0.33218567  0.127119532
#>    21   41 0.449852970 0.99301774 0.665094146 0.92758251  0.700747419
#>    66   16 0.064687161 0.51903636 0.033782071 0.32910251  0.029200203
#>    13   78 0.886276338 0.94994035 0.795554876 0.87804735  0.843337247
#>    42   38 0.379801397 0.26304469 0.457780416 0.30395176  0.401617802
#>     4   55 0.664216228 0.44892703 0.626505153 0.35913803  0.689467034
#>    19   26 0.900917128 0.72823690 0.902980245 0.71761406  0.971085161
#>     4   32 0.664216228 0.44892703 0.891671680 0.57587259  0.790162361
#>    36   16 0.000000000 0.77575771 0.033782071 0.32910251 -0.499570821
#>    42   20 0.379801397 0.26304469 0.479962892 0.24615722  0.430871588
#>    77   54 0.373287922 0.36956137 0.143674555 0.33218567  0.257669511
#>    19  100 0.900917128 0.72823690 0.937106278 0.79827905  0.883938198
#>    54   74 0.143674555 0.33218567 0.139257838 0.27630199  0.212448982
#>    91   42 0.331096759 0.25110430 0.379801397 0.26304469  0.341454707
#>   100   88 0.937106278 0.79827905 0.932702891 0.81776290  0.884563422
#>    57   53 0.945052348 0.86943517 0.941993173 0.94921684  0.941299395
#>    68   94 0.901345061 0.91073078 0.882722195 0.86833393  0.913690539
#>    53   28 0.941993173 0.94921684 0.951100382 0.99270841  0.914499578
#>    53   13 0.941993173 0.94921684 0.886276338 0.94994035  0.913985558
#>    91   77 0.331096759 0.25110430 0.373287922 0.36956137  0.281059555
#>    86   28 1.000000000 0.97250373 0.951100382 0.99270841  0.968812774
#>    26   32 0.902980245 0.71761406 0.891671680 0.57587259  0.962423865
#>    90   51 0.750211405 0.68281735 0.674265294 0.55499947  0.781803042
#>    38   20 0.457780416 0.30395176 0.479962892 0.24615722  0.537402735
#>    91   22 0.331096759 0.25110430 0.260993014 0.23948033  0.314005666
#>    55   52 0.626505153 0.35913803 0.568239211 0.22790339  0.627674398
#>    57   86 0.945052348 0.86943517 1.000000000 0.97250373  0.990954924
#>    57  100 0.945052348 0.86943517 0.937106278 0.79827905  1.187557080
#>    56   86 0.984325183 0.54753006 1.000000000 0.97250373  1.352695904
#>    55   38 0.626505153 0.35913803 0.457780416 0.30395176  0.547309754
#>    15   22 0.182101540 0.17577138 0.260993014 0.23948033  0.194901750
#>    97   15 0.095099748 0.21096931 0.182101540 0.17577138  0.149969757
#>    74   16 0.139257838 0.27630199 0.033782071 0.32910251  0.072177675
#>    97   62 0.095099748 0.21096931 0.004990548 0.21004585  0.050534872
#>    97   16 0.095099748 0.21096931 0.033782071 0.32910251  0.049514485
#>    56   32 0.984325183 0.54753006 0.891671680 0.57587259  0.911211838
#>    12    8 0.684449752 0.22502027 0.737985514 0.26983202  0.672957381
#>   100   26 0.937106278 0.79827905 0.902980245 0.71761406  1.071671658
#>    32   64 0.891671680 0.57587259 0.730674334 0.33108805  0.814273011
#>    55   12 0.626505153 0.35913803 0.684449752 0.22502027  0.665471854
#>    61   12 0.655550400 0.15435393 0.684449752 0.22502027  0.625879763
#>    48   69 0.648421934 0.10472636 0.760061029 0.12228873  0.702869105
#>    72   52 0.507439193 0.05382380 0.568239211 0.22790339  0.506480416
#>    49   58 0.859980528 0.21244083 0.938860116 0.37429399  0.841189323
#>    48   61 0.648421934 0.10472636 0.655550400 0.15435393  0.569941088
#>    15   47 0.182101540 0.17577138 0.051338303 0.09704640  0.152588084
#>    98   31 0.249517059 0.04126542 0.429181597 0.07756247  0.324505311
#>    31   72 0.429181597 0.07756247 0.507439193 0.05382380  0.493889332
#>    72   48 0.507439193 0.05382380 0.648421934 0.10472636  0.558270131
#>    98   72 0.249517059 0.04126542 0.507439193 0.05382380  0.390141694
#>     8   49 0.737985514 0.26983202 0.859980528 0.21244083  0.832397551
#>    25   69 0.808289559 0.02300336 0.760061029 0.12228873  0.715887827
#>    98   46 0.249517059 0.04126542 0.042431837 0.01650487  0.413212437
#>    98   47 0.249517059 0.04126542 0.051338303 0.09704640  0.143922697
#>    47   62 0.051338303 0.09704640 0.004990548 0.21004585 -0.153692277
#>    24   49 0.893608991 0.06287428 0.859980528 0.21244083  0.843565049
#>    25   24 0.808289559 0.02300336 0.893608991 0.06287428  0.827281545
#>    72   25 0.507439193 0.05382380 0.808289559 0.02300336  0.643457932
#>    98   37 0.249517059 0.04126542 0.830078009 0.00832426  0.453575582
#>    24   81 0.893608991 0.06287428 0.931806982 0.06392915  0.910448379
#>    81   58 0.931806982 0.06392915 0.938860116 0.37429399  1.057481788
#>    37   25 0.830078009 0.00832426 0.808289559 0.02300336  0.626186638
#>    85   56 0.980812802 0.00000000 0.984325183 0.54753006  1.692022160
#>    85   81 0.980812802 0.00000000 0.931806982 0.06392915  0.914462049
#>    37   85 0.830078009 0.00832426 0.980812802 0.00000000  0.904441735
#>    37   46 0.830078009 0.00832426 0.042431837 0.01650487  0.327468318
#>     1   99 0.344803363 0.84680901 0.318085050 0.83410145  0.312224114
#>    50    1 0.405913711 0.87201231 0.344803363 0.84680901  0.360260343
#>    45   14 0.240357038 1.00000000 0.269130021 0.99115797  0.236368248
#>     1   65 0.344803363 0.84680901 0.350738059 0.73668302  0.354439971
#>    89   50 0.464730460 0.95580398 0.405913711 0.87201231  0.391447485
#>    99   59 0.318085050 0.83410145 0.163880024 0.81531479  0.242312264
#>    10   67 0.073201781 0.88280553 0.164802363 0.98364114  0.138135092
#>    21   34 0.449852970 0.99301774 0.319364664 0.92745331  0.359300951
#>    79   75 0.478250328 0.92981364 0.463449311 0.92122834  0.463294906
#>    41   89 0.665094146 0.92758251 0.464730460 0.95580398  0.576217341
#>     2   50 0.550956198 0.89005271 0.405913711 0.87201231  0.483610718
#>    10   17 0.073201781 0.88280553 0.202246247 0.88123708  0.137635206
#>    65   44 0.350738059 0.73668302 0.440802661 0.66171352  0.377201054
#>     2   41 0.550956198 0.89005271 0.665094146 0.92758251  0.602945829
#>     2   43 0.550956198 0.89005271 0.570689686 0.87103967  0.493039789
#>    59   60 0.163880024 0.81531479 0.016906306 0.81584939  0.090377300
#>    73   40 0.358015069 0.60447089 0.331518855 0.63010726  0.383007745
#>    92   35 0.093664681 0.66284925 0.158926877 0.56442975  0.154825071
#>     2   87 0.550956198 0.89005271 0.590170156 0.82743770  0.487664495
#>    92   59 0.093664681 0.66284925 0.163880024 0.81531479  0.219582314
#>    36   92 0.000000000 0.77575771 0.093664681 0.66284925  0.001481859
#>    63   44 0.565504476 0.68829859 0.440802661 0.66171352  0.481861060
#>    92   40 0.093664681 0.66284925 0.331518855 0.63010726  0.216285579
#>     7   36 0.080060133 0.73502438 0.000000000 0.77575771  0.051381121
#>    40   84 0.331518855 0.63010726 0.196093551 0.51396227  0.247661889
#>    44   39 0.440802661 0.66171352 0.511248648 0.58927462  0.433354996
#>    40    9 0.331518855 0.63010726 0.316779684 0.47303579  0.277610805
#>    30   63 0.600933809 0.64549187 0.565504476 0.68829859  0.591730297
#>     5   43 0.684431950 0.90207472 0.570689686 0.87103967  0.629259463
#>    73   96 0.358015069 0.60447089 0.388504122 0.45311624  0.370752967
#>    51   71 0.674265294 0.55499947 0.620659969 0.68149268  0.685922505
#>     5   78 0.684431950 0.90207472 0.795554876 0.87804735  0.745215892
#>    80   87 0.749206139 0.85235847 0.590170156 0.82743770  0.672822659
#>    35   11 0.158926877 0.56442975 0.126976444 0.43778443  0.129774448
#>    35   66 0.158926877 0.56442975 0.064687161 0.51903636  0.094392143
#>    66   36 0.064687161 0.51903636 0.000000000 0.77575771 -0.101374112
#>     3   23 0.761596075 0.84080656 0.675419183 0.72011230  0.703644843
#>    90   23 0.750211405 0.68281735 0.675419183 0.72011230  0.740435879
#>    33   27 0.551179583 0.42633543 0.474154623 0.43693288  0.516773820
#>     9   95 0.316779684 0.47303579 0.141701822 0.34328150  0.231126433
#>    51   30 0.674265294 0.55499947 0.600933809 0.64549187  0.673350530
#>     5   41 0.684431950 0.90207472 0.665094146 0.92758251  0.759187800
#>    78    3 0.795554876 0.87804735 0.761596075 0.84080656  0.772522759
#>    78   83 0.795554876 0.87804735 0.774688881 0.82321841  0.790438211
#>    51   93 0.674265294 0.55499947 0.544106450 0.58969567  0.610842699
#>    70    4 0.528794122 0.50917568 0.664216228 0.44892703  0.599839888
#>    83   18 0.774688881 0.82321841 0.832330316 0.80257071  0.780908482
#>    32   90 0.891671680 0.57587259 0.750211405 0.68281735  0.837691719
#>    70   82 0.528794122 0.50917568 0.619497870 0.41462239  0.596041182
#>    27   29 0.474154623 0.43693288 0.455503619 0.33626992  0.430603859
#>    54   95 0.143674555 0.33218567 0.141701822 0.34328150  0.257303139
#>     9   77 0.316779684 0.47303579 0.373287922 0.36956137  0.255670693
#>    11   76 0.126976444 0.43778443 0.110336345 0.33461694  0.067551113
#>    94   18 0.882722195 0.86833393 0.832330316 0.80257071  0.836710599
#>    29   42 0.455503619 0.33626992 0.379801397 0.26304469  0.400154910
#>    28   41 0.951100382 0.99270841 0.665094146 0.92758251  0.700747419
#>    11   16 0.126976444 0.43778443 0.033782071 0.32910251  0.029200203
#>    68   78 0.901345061 0.91073078 0.795554876 0.87804735  0.843337247
#>    55   29 0.626505153 0.35913803 0.455503619 0.33626992  0.542003573
#>    57   88 0.945052348 0.86943517 0.932702891 0.81776290  0.914229938
#>    64   55 0.730674334 0.33108805 0.626505153 0.35913803  0.689467034
#>    55   82 0.626505153 0.35913803 0.619497870 0.41462239  0.671589475
#>    55   33 0.626505153 0.35913803 0.551179583 0.42633543  0.578617256
#>    51   32 0.674265294 0.55499947 0.891671680 0.57587259  0.790162361
#>    74   76 0.139257838 0.27630199 0.110336345 0.33461694  0.124974131
#>    22   54 0.260993014 0.23948033 0.143674555 0.33218567  0.257669511
#>    18  100 0.832330316 0.80257071 0.937106278 0.79827905  0.883938198
#>    22   74 0.260993014 0.23948033 0.139257838 0.27630199  0.212448982
#>    18   88 0.832330316 0.80257071 0.932702891 0.81776290  0.884563422
#>    68   53 0.901345061 0.91073078 0.941993173 0.94921684  0.941299395
#>    57   94 0.945052348 0.86943517 0.882722195 0.86833393  0.913690539
#>    13   28 0.886276338 0.94994035 0.951100382 0.99270841  0.914499578
#>    68   13 0.901345061 0.91073078 0.886276338 0.94994035  0.913985558
#>    22   77 0.260993014 0.23948033 0.373287922 0.36956137  0.281059555
#>    28   21 0.951100382 0.99270841 0.449852970 0.99301774  0.707183213
#>    52   20 0.568239211 0.22790339 0.479962892 0.24615722  0.537402735
#>    31   42 0.429181597 0.07756247 0.379801397 0.26304469  0.416194719
#>    98   22 0.249517059 0.04126542 0.260993014 0.23948033  0.314005666
#>    64    4 0.730674334 0.33108805 0.664216228 0.44892703  0.810627297
#>    12   52 0.684449752 0.22502027 0.568239211 0.22790339  0.627674398
#>    53   86 0.941993173 0.94921684 1.000000000 0.97250373  0.990954924
#>    86  100 1.000000000 0.97250373 0.937106278 0.79827905  1.187557080
#>    52   38 0.568239211 0.22790339 0.457780416 0.30395176  0.547309754
#>    31   91 0.429181597 0.07756247 0.331096759 0.25110430  0.378423186
#>    74   15 0.139257838 0.27630199 0.182101540 0.17577138  0.149969757
#>    98   15 0.249517059 0.04126542 0.182101540 0.17577138  0.276868048
#>    62   16 0.004990548 0.21004585 0.033782071 0.32910251  0.049514485
#>    62   36 0.004990548 0.21004585 0.000000000 0.77575771 -0.872194413
#>    58   32 0.938860116 0.37429399 0.891671680 0.57587259  0.911211838
#>    64    8 0.730674334 0.33108805 0.737985514 0.26983202  0.672957381
#>    56   26 0.984325183 0.54753006 0.902980245 0.71761406  1.071671658
#>    58   64 0.938860116 0.37429399 0.730674334 0.33108805  0.814273011
#>    64   12 0.730674334 0.33108805 0.684449752 0.22502027  0.665471854
#>    61   69 0.655550400 0.15435393 0.760061029 0.12228873  0.702869105
#>    69   12 0.760061029 0.12228873 0.684449752 0.22502027  0.753635394
#>    52   61 0.568239211 0.22790339 0.655550400 0.15435393  0.569941088
#>    47   97 0.051338303 0.09704640 0.095099748 0.21096931  0.116023791
#>    20   72 0.479962892 0.24615722 0.507439193 0.05382380  0.493889332
#>    52   48 0.568239211 0.22790339 0.648421934 0.10472636  0.558270131
#>    64   49 0.730674334 0.33108805 0.859980528 0.21244083  0.832397551
#>    69    8 0.760061029 0.12228873 0.737985514 0.26983202  0.779954675
#>    46   47 0.042431837 0.01650487 0.051338303 0.09704640  0.143922697
#>    46   62 0.042431837 0.01650487 0.004990548 0.21004585 -0.153692277
#>    69   49 0.760061029 0.12228873 0.859980528 0.21244083  0.843565049
#>    69   24 0.760061029 0.12228873 0.893608991 0.06287428  0.827281545
#>    48   25 0.648421934 0.10472636 0.808289559 0.02300336  0.643457932
#>    72   37 0.507439193 0.05382380 0.830078009 0.00832426  0.453575582
#>    49   81 0.859980528 0.21244083 0.931806982 0.06392915  0.910448379
#>    58   56 0.938860116 0.37429399 0.984325183 0.54753006  1.692022160
#>    58   85 0.938860116 0.37429399 0.980812802 0.00000000  1.192814019
#>    24   85 0.893608991 0.06287428 0.980812802 0.00000000  0.904441735
#>    85   46 0.980812802 0.00000000 0.042431837 0.01650487  0.327468318
#>    17   99 0.202246247 0.88123708 0.318085050 0.83410145  0.269841161
#>    45   17 0.240357038 1.00000000 0.202246247 0.88123708  0.236368248
#>    21   14 0.449852970 0.99301774 0.269130021 0.99115797  0.359300951
#>    79   89 0.478250328 0.92981364 0.464730460 0.95580398  0.463294906
#>    75    2 0.463449311 0.92122834 0.550956198 0.89005271  0.497758310
#>    10   59 0.073201781 0.88280553 0.163880024 0.81531479  0.137635206
#>    65   40 0.350738059 0.73668302 0.331518855 0.63010726  0.377201054
#>    73   44 0.358015069 0.60447089 0.440802661 0.66171352  0.383007745
#>    63   87 0.565504476 0.68829859 0.590170156 0.82743770  0.481861060
#>     5   80 0.684431950 0.90207472 0.749206139 0.85235847  0.745215892
#>    90   83 0.750211405 0.68281735 0.774688881 0.82321841  0.740435879
#>    33   70 0.551179583 0.42633543 0.528794122 0.50917568  0.516773820
#>    77   27 0.373287922 0.36956137 0.474154623 0.43693288  0.423498649
#>     4   51 0.664216228 0.44892703 0.674265294 0.55499947  0.609228008
#>    95   76 0.141701822 0.34328150 0.110336345 0.33461694  0.127119532
#>    19   18 0.900917128 0.72823690 0.832330316 0.80257071  0.820378492
#>    68   57 0.901345061 0.91073078 0.945052348 0.86943517  0.941299395
#>    28   45 0.951100382 0.99270841 0.240357038 1.00000000  0.707183213
#>    31   20 0.429181597 0.07756247 0.479962892 0.24615722  0.416194719
#>    98   91 0.249517059 0.04126542 0.331096759 0.25110430  0.314005666
#>   100   56 0.937106278 0.79827905 0.984325183 0.54753006  1.352695904
#>    74   97 0.139257838 0.27630199 0.095099748 0.21096931  0.149969757
#>    37   24 0.830078009 0.00832426 0.893608991 0.06287428  0.845465301
#>    85   86 0.980812802 0.00000000 1.000000000 0.97250373 16.952955463
#>            my1           mx2           my2 bp1 bp2
#>   8.808665e-01   0.269841161  8.814475e-01   0   0
#>   8.960192e-01   0.312224114  8.808665e-01   0   0
#>   7.946716e-01   0.456698355  7.723941e-01   0   0
#>   9.261499e-01   0.269841161  8.814475e-01   0   0
#>   9.410129e-01   0.483610718  8.394200e-01   0   0
#>   9.357837e-01   0.252201094  9.261499e-01   0   0
#>   7.921054e-01   0.249249937  7.568475e-01   0   0
#>   7.568475e-01   0.227001131  7.039761e-01   0   0
#>   9.447055e-01   0.360260343  8.960192e-01   0   0
#>   7.763065e-01   0.456698355  7.723941e-01   0   0
#>   8.137934e-01   0.137635206  8.747146e-01   0   0
#>   9.158426e-01   0.213095652  9.432518e-01   0   0
#>   1.010604e+00   0.252201094  9.261499e-01   0   0
#>   9.478427e-01   0.391447485  9.447055e-01   0   0
#>   9.385456e-01   0.396708052  9.410129e-01   0   0
#>   8.791309e-01   0.556870903  9.872227e-01   0   0
#>   1.325283e+00   0.707183213  1.186031e+01   0   0
#>   1.021955e+00   0.390835629  9.478427e-01   0   0
#>   9.432518e-01  -2.524275090  1.358601e+01   0   1
#>   6.768900e-01   0.383007745  6.568124e-01   0   0
#>   7.162015e-01   0.090366178  8.081628e-01   0   0
#>   9.242652e-01   0.629259463  8.803317e-01   0   0
#>   8.101946e-01   0.631515684  8.720628e-01   0   0
#>   8.112205e-01  -8.659422454  8.167888e+00   0   1
#>   6.325573e-01   0.094392143  5.778875e-01   0   0
#>   8.068282e-01   0.480280805  7.763065e-01   0   0
#>   8.081628e-01   0.051381121  7.777012e-01   0   0
#>   6.972611e-01   0.227001131  7.039761e-01   0   0
#>   6.816824e-01   0.154825071  6.325573e-01   0   0
#>   7.748822e-01   0.480280805  7.763065e-01   0   0
#>   6.558476e-01   0.526965666  6.450387e-01   0   0
#>   6.733119e-01   0.247661889  5.908591e-01   0   0
#>   7.777012e-01 -21.125675991  9.707880e+00   0   1
#>   6.831844e-01   0.178455172  7.162015e-01   0   0
#>   5.908591e-01   0.130169267  5.043319e-01   0   0
#>   5.839974e-01   0.428993807  5.400207e-01   0   0
#>   5.552679e-01   0.370752967  5.282886e-01   0   0
#>   1.010472e+00   0.556870903  9.872227e-01   0   0
#>   7.553352e-01   0.601567567  7.536614e-01   0   0
#>   7.536614e-01   0.591730297  6.739396e-01   0   0
#>   6.399073e-01   0.526965666  6.450387e-01   0   0
#>   6.450387e-01   0.528170581  5.510102e-01   0   0
#>   5.559386e-01   0.229536544  4.141747e-01   0   0
#>   5.282886e-01   0.340325194  4.187272e-01   0   0
#>   5.043319e-01   0.229536544  4.141747e-01   0   0
#>   8.071551e-01   0.609579683  7.553352e-01   0   0
#>   6.345446e-01   0.685792363  6.472724e-01   0   0
#>   5.400207e-01   0.438574038  5.313843e-01   0   0
#>   6.472724e-01   0.609579683  7.553352e-01   0   0
#>   5.373765e-01   0.528170581  5.510102e-01   0   0
#>   5.313843e-01   0.528170581  5.510102e-01   0   0
#>   9.142142e-01   0.772522759  8.649463e-01   0   0
#>   8.198947e-01   0.631515684  8.720628e-01   0   0
#>   4.141747e-01   0.231126433  4.056143e-01   0   0
#>   5.044315e-01   0.029200203  4.273293e-01   0   0
#>   5.778875e-01  -0.101374112  6.137036e-01   0   0
#>   6.137036e-01   0.001481859  6.816824e-01   0   0
#>   7.910716e-01   0.674818944  8.071551e-01   0   0
#>   7.876094e-01   0.790438211  8.486097e-01   0   0
#>   7.568558e-01   0.708493819  7.876094e-01   0   0
#>   4.614828e-01   0.444734035  5.159688e-01   0   0
#>   5.159688e-01   0.438574038  5.313843e-01   0   0
#>   4.056143e-01   0.112700728  3.871613e-01   0   0
#>   4.187272e-01   0.423498649  4.035804e-01   0   0
#>   6.292169e-01   0.591730297  6.739396e-01   0   0
#>   9.788322e-01   0.769017785  1.004721e+00   0   0
#>   8.649463e-01   0.703644843  7.910716e-01   0   0
#>   8.399473e-01   0.836710599  8.514026e-01   0   0
#>   5.785630e-01   0.550611521  6.399073e-01   0   0
#>   4.865469e-01   0.671589475  3.930165e-01   0   0
#>   3.790777e-01   0.542003573  3.402323e-01   0   0
#>   4.035804e-01   0.444734035  5.159688e-01   0   0
#>   7.497999e-01   0.820378492  7.227339e-01   0   0
#>   7.079487e-01   0.820378492  7.227339e-01   0   0
#>   6.515011e-01   0.824834448  7.079487e-01   0   0
#>   5.076487e-01   0.599863692  5.373765e-01   0   0
#>   4.829028e-01   0.578617256  3.812748e-01   0   0
#>   3.929427e-01   0.400154910  3.177468e-01   0   0
#>   3.262562e-01   0.401617802  3.162344e-01   0   0
#>   3.581110e-01   0.255670693  3.724966e-01   0   0
#>   3.944436e-01   0.075714373  2.811167e-01   0   0
#>   1.004721e+00   0.745923855  1.233186e+00   0   0
#>   9.109198e-01   0.836710599  8.514026e-01   0   0
#>   8.171082e-01   0.914229938  8.494897e-01   0   0
#>   3.177468e-01   0.341454707  3.141573e-01   0   0
#>   3.349656e-01   0.124974131  3.055473e-01   0   0
#>   1.431582e+00   0.576217341  1.021955e+00   0   0
#>   4.273293e-01  -0.499570821  5.133683e-01   0   0
#>   9.109380e-01   0.769017785  1.004721e+00   0   0
#>   3.162344e-01   0.430871588  2.604695e-01   0   0
#>   3.855080e-01   0.671589475  3.930165e-01   0   0
#>   7.363528e-01   0.824834448  7.079487e-01   0   0
#>   4.905074e-01   0.810627297  4.538392e-01   0   0
#>   5.133683e-01  -0.872194413  4.851855e-01   0   0
#>   2.604695e-01   0.416194719  1.734193e-01   0   0
#>   3.558603e-01   0.257303139  3.581110e-01   0   0
#>   7.813796e-01   0.971085161  7.363528e-01   0   0
#>   2.986338e-01   0.124974131  3.055473e-01   0   0
#>   3.141573e-01   0.378423186  1.633635e-01   0   0
#>   7.966438e-01   1.013458941  8.257744e-01   0   0
#>   9.092408e-01   0.990954924  9.111448e-01   0   0
#>   8.800195e-01   0.843342859  9.109198e-01   0   0
#>   9.776734e-01   0.968812774  9.663001e-01   0   0
#>   9.380890e-01   0.914499578  9.776734e-01   0   0
#>   3.356684e-01   0.341454707  3.141573e-01   0   0
#>   9.663001e-01   8.186311818  1.843418e+01   0   1
#>   6.415496e-01   0.837691719  6.515011e-01   0   0
#>   5.775749e-01   0.685922505  6.345446e-01   0   0
#>   3.013578e-01   0.430871588  2.604695e-01   0   0
#>   1.369714e-01   0.281059555  3.356684e-01   0   0
#>   2.800670e-01   0.547309754  3.157476e-01   0   0
#>   9.111448e-01   1.187557080  8.063327e-01   0   0
#>   8.063327e-01   1.013458941  8.257744e-01   0   0
#>   7.467189e-01  16.952955463  1.713157e-01   0   0
#>   3.157476e-01   0.543872609  3.262562e-01   0   0
#>   2.406213e-01   0.276868048  1.391216e-01   0   0
#>   2.214724e-01   0.116023791  1.375652e-01   0   0
#>   2.740517e-01   0.075714373  2.811167e-01   0   0
#>   1.627216e-01   0.049514485  2.622883e-01   0   0
#>   2.622883e-01   0.072177675  2.740517e-01   0   0
#>   4.741343e-01   0.962423865  6.415496e-01   0   0
#>   2.931349e-01   0.753635394  1.967504e-01   0   0
#>   6.937988e-01   0.971085161  7.363528e-01   0   0
#>   4.514414e-01   0.810627297  4.538392e-01   0   0
#>   2.963972e-01   0.627674398  2.800670e-01   0   0
#>   2.077303e-01   0.717593695  1.702234e-01   0   0
#>   1.222314e-01   0.715887827  3.947495e-02   0   0
#>   1.518161e-01   0.558270131  1.337278e-01   0   0
#>   3.217464e-01   1.057481788  2.163357e-01   0   0
#>   1.413250e-01   0.702869105  1.222314e-01   0   0
#>   7.683140e-02   0.116023791  1.375652e-01   0   0
#>   1.328894e-01   0.390141694 -1.920001e-01   0   0
#>   1.500174e-01   0.390141694 -1.920001e-01   0   0
#>   1.337278e-01   0.643457932 -1.022134e-01   0   0
#>  -1.920001e-01   0.453575582 -1.494796e+00   0   0
#>   3.121648e-01   0.779954675  2.006883e-01   0   0
#>   3.947495e-02   0.827281545  9.358519e-02   0   0
#>  -2.206164e+00   0.143922697  4.604499e-02   0   0
#>   4.604499e-02   0.152588084  7.683140e-02   0   0
#>   7.895594e-02   0.050534872  1.627216e-01   0   0
#>   1.301862e-01   0.910448379  1.452242e-01   0   0
#>   9.358519e-02   0.845465301  5.467393e-02   0   0
#>  -1.022134e-01   0.626186638 -2.708053e-01   0   0
#>  -1.494796e+00   0.413212437 -2.206164e+00   0   0
#>   1.452242e-01   0.914462049 -1.144971e-04   0   0
#>   2.163357e-01   1.192814019  2.132603e-01   0   0
#>  -2.708053e-01   0.845465301  5.467393e-02   0   0
#>   2.692139e-01  16.952955463  1.713157e-01   0   0
#>  -1.144971e-04   1.192814019  2.132603e-01   0   0
#>  -1.401222e-02   0.327468318 -1.046178e+01   0   0
#>  -1.046178e+01   0.413212437 -2.206164e+00   0   0
#>   8.808665e-01   0.354439971  7.921054e-01   0   0
#>   8.960192e-01   0.402058411  7.946716e-01   0   0
#>   9.357837e-01   0.356062677  1.325283e+00   0   0
#>   7.921054e-01   0.402058411  7.946716e-01   0   0
#>   9.447055e-01   0.396708052  9.410129e-01   0   0
#>   8.137934e-01   0.249249937  7.568475e-01   0   0
#>   9.158426e-01  -5.295261060  5.851621e+00   0   1
#>   1.010604e+00   0.390835629  9.478427e-01   0   0
#>   9.385456e-01   0.497758310  8.791309e-01   0   0
#>   1.021955e+00   0.574599943  1.010472e+00   0   0
#>   8.394200e-01   0.487664495  8.068282e-01   0   0
#>   8.747146e-01   0.138135092  9.158426e-01   0   0
#>   6.768900e-01   0.456698355  7.723941e-01   0   0
#>   9.242652e-01   0.574599943  1.010472e+00   0   0
#>   8.101946e-01   0.602945829  9.242652e-01   0   0
#>   8.112205e-01   0.090366178  8.081628e-01   0   0
#>   6.568124e-01   0.284758423  5.552679e-01   0   0
#>   6.325573e-01   0.216285579  6.733119e-01   0   0
#>   8.068282e-01   0.493039789  8.101946e-01   0   0
#>   6.972611e-01   0.178455172  7.162015e-01   0   0
#>   6.816824e-01   0.003292476  6.831844e-01   0   0
#>   7.748822e-01   0.507237933  6.558476e-01   0   0
#>   6.733119e-01   0.219582314  6.972611e-01   0   0
#>   7.777012e-01   0.003292476  6.831844e-01   0   0
#>   5.908591e-01   0.277610805  5.559386e-01   0   0
#>   5.839974e-01   0.507237933  6.558476e-01   0   0
#>   5.559386e-01   0.284758423  5.552679e-01   0   0
#>   6.739396e-01   0.550611521  6.399073e-01   0   0
#>   8.803317e-01   0.631515684  8.720628e-01   0   0
#>   5.282886e-01   0.428993807  5.400207e-01   0   0
#>   6.345446e-01   0.673350530  6.292169e-01   0   0
#>   9.142142e-01   0.759187800  9.788322e-01   0   0
#>   8.198947e-01   0.674818944  8.071551e-01   0   0
#>   5.044315e-01   0.130169267  5.043319e-01   0   0
#>   5.778875e-01   0.129774448  5.044315e-01   0   0
#>   6.137036e-01  -0.499570821  5.133683e-01   0   0
#>   7.910716e-01   0.708493819  7.876094e-01   0   0
#>   7.568558e-01   0.685792363  6.472724e-01   0   0
#>   4.614828e-01   0.505436140  3.790777e-01   0   0
#>   4.056143e-01   0.255670693  3.724966e-01   0   0
#>   6.292169e-01   0.610842699  5.785630e-01   0   0
#>   9.788322e-01   0.629259463  8.803317e-01   0   0
#>   8.649463e-01   0.790438211  8.486097e-01   0   0
#>   8.486097e-01   0.813200134  8.399473e-01   0   0
#>   5.785630e-01   0.599863692  5.373765e-01   0   0
#>   4.865469e-01   0.609228008  5.076487e-01   0   0
#>   7.497999e-01   0.813200134  8.399473e-01   0   0
#>   6.515011e-01   0.781803042  5.775749e-01   0   0
#>   4.829028e-01   0.599839888  4.865469e-01   0   0
#>   3.929427e-01   0.505436140  3.790777e-01   0   0
#>   3.581110e-01   0.127119532  3.349656e-01   0   0
#>   3.724966e-01   0.340325194  4.187272e-01   0   0
#>   3.944436e-01   0.112700728  3.871613e-01   0   0
#>   8.514026e-01   0.881465964  8.171082e-01   0   0
#>   3.177468e-01   0.401617802  3.162344e-01   0   0
#>   1.431582e+00   0.745923855  1.233186e+00   0   0
#>   4.273293e-01   0.067551113  3.944436e-01   0   0
#>   9.109380e-01   0.843342859  9.109198e-01   0   0
#>   3.402323e-01   0.543872609  3.262562e-01   0   0
#>   8.494897e-01   1.013458941  8.257744e-01   0   0
#>   3.855080e-01   0.665471854  2.963972e-01   0   0
#>   3.930165e-01   0.578617256  3.812748e-01   0   0
#>   3.812748e-01   0.542003573  3.402323e-01   0   0
#>   4.905074e-01   0.781803042  5.775749e-01   0   0
#>   3.055473e-01   0.075714373  2.811167e-01   0   0
#>   3.558603e-01   0.212448982  2.986338e-01   0   0
#>   7.813796e-01   0.884563422  7.966438e-01   0   0
#>   2.986338e-01   0.194901750  2.406213e-01   0   0
#>   7.966438e-01   0.881465964  8.171082e-01   0   0
#>   9.092408e-01   0.913985558  9.380890e-01   0   0
#>   8.800195e-01   0.914229938  8.494897e-01   0   0
#>   9.776734e-01   0.745923855  1.233186e+00   0   0
#>   9.380890e-01   0.843337247  9.109380e-01   0   0
#>   3.356684e-01   0.257669511  3.558603e-01   0   0
#>   1.186031e+01   0.700747419  1.431582e+00   0   0
#>   3.013578e-01   0.506480416  1.518161e-01   0   0
#>   1.734193e-01   0.378423186  1.633635e-01   0   0
#>   1.369714e-01   0.276868048  1.391216e-01   0   0
#>   4.538392e-01   0.689467034  3.855080e-01   0   0
#>   2.800670e-01   0.625879763  2.077303e-01   0   0
#>   9.111448e-01   0.968812774  9.663001e-01   0   0
#>   8.063327e-01   1.352695904  7.467189e-01   0   0
#>   3.157476e-01   0.537402735  3.013578e-01   0   0
#>   1.633635e-01   0.324505311  1.328894e-01   0   0
#>   2.214724e-01   0.194901750  2.406213e-01   0   0
#>   1.391216e-01   0.152588084  7.683140e-02   0   0
#>   2.622883e-01  -0.872194413  4.851855e-01   0   0
#>   4.851855e-01  -2.639741190  4.695927e-01   0   1
#>   4.741343e-01   0.814273011  4.514414e-01   0   0
#>   2.931349e-01   0.832397551  3.121648e-01   0   0
#>   6.937988e-01   0.962423865  6.415496e-01   0   0
#>   4.514414e-01   0.841189323  3.217464e-01   0   0
#>   2.963972e-01   0.672957381  2.931349e-01   0   0
#>   1.222314e-01   0.717593695  1.702234e-01   0   0
#>   1.967504e-01   0.717593695  1.702234e-01   0   0
#>   1.413250e-01   0.625879763  2.077303e-01   0   0
#>   1.375652e-01   0.050534872  1.627216e-01   0   0
#>   1.500174e-01   0.506480416  1.518161e-01   0   0
#>   1.337278e-01   0.569941088  1.413250e-01   0   0
#>   3.121648e-01   0.841189323  3.217464e-01   0   0
#>   2.006883e-01   0.753635394  1.967504e-01   0   0
#>   4.604499e-02  -0.153692277  7.895594e-02   0   0
#>   7.895594e-02  -5.134165211 -8.845368e-01   0   1
#>   1.301862e-01   0.779954675  2.006883e-01   0   0
#>   9.358519e-02   0.843565049  1.301862e-01   0   0
#>  -1.022134e-01   0.715887827  3.947495e-02   0   0
#>  -1.494796e+00   0.626186638 -2.708053e-01   0   0
#>   1.452242e-01   1.057481788  2.163357e-01   0   0
#>   2.692139e-01   0.911211838  4.741343e-01   0   0
#>   2.132603e-01   1.692022160  2.692139e-01   0   0
#>  -1.401222e-02   0.914462049 -1.144971e-04   0   0
#>  -1.046178e+01   0.308730483 -1.152712e+01   0   1
#>   8.814475e-01   0.242312264  8.137934e-01   0   0
#>   9.357837e-01   0.213095652  9.432518e-01   0   0
#>   1.010604e+00   0.356062677  1.325283e+00   0   0
#>   9.385456e-01   0.556870903  9.872227e-01   0   0
#>   8.791309e-01   0.483610718  8.394200e-01   0   0
#>   8.747146e-01   0.090377300  8.112205e-01   0   0
#>   6.768900e-01   0.227001131  7.039761e-01   0   0
#>   6.568124e-01   0.433354996  5.839974e-01   0   0
#>   7.748822e-01   0.601567567  7.536614e-01   0   0
#>   9.142142e-01   0.672822659  8.198947e-01   0   0
#>   7.568558e-01   0.780908482  7.497999e-01   0   0
#>   4.614828e-01   0.596041182  4.829028e-01   0   0
#>   4.035804e-01   0.430603859  3.929427e-01   0   0
#>   5.076487e-01   0.790162361  4.905074e-01   0   0
#>   3.349656e-01   0.112700728  3.871613e-01   0   0
#>   7.227339e-01   0.883938198  7.813796e-01   0   0
#>   9.092408e-01   0.913690539  8.800195e-01   0   0
#>   1.186031e+01   0.721616024  1.326714e+01   0   1
#>   1.734193e-01   0.493889332  1.500174e-01   0   0
#>   1.369714e-01   0.324505311  1.328894e-01   0   0
#>   7.467189e-01   1.071671658  6.937988e-01   0   0
#>   2.214724e-01   0.072177675  2.740517e-01   0   0
#>   5.467393e-02   0.904441735 -1.401222e-02   0   0
#>   1.713157e-01  17.980829042  1.510361e-01   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.344803363 0.84680901
#>   [2,] 0.550956198 0.89005271
#>   [3,] 0.761596075 0.84080656
#>   [4,] 0.664216228 0.44892703
#>   [5,] 0.684431950 0.90207472
#>   [6,] 0.087447929 0.64344374
#>   [7,] 0.080060133 0.73502438
#>   [8,] 0.737985514 0.26983202
#>   [9,] 0.316779684 0.47303579
#>  [10,] 0.073201781 0.88280553
#>  [11,] 0.126976444 0.43778443
#>  [12,] 0.684449752 0.22502027
#>  [13,] 0.886276338 0.94994035
#>  [14,] 0.269130021 0.99115797
#>  [15,] 0.182101540 0.17577138
#>  [16,] 0.033782071 0.32910251
#>  [17,] 0.202246247 0.88123708
#>  [18,] 0.832330316 0.80257071
#>  [19,] 0.900917128 0.72823690
#>  [20,] 0.479962892 0.24615722
#>  [21,] 0.449852970 0.99301774
#>  [22,] 0.260993014 0.23948033
#>  [23,] 0.675419183 0.72011230
#>  [24,] 0.893608991 0.06287428
#>  [25,] 0.808289559 0.02300336
#>  [26,] 0.902980245 0.71761406
#>  [27,] 0.474154623 0.43693288
#>  [28,] 0.951100382 0.99270841
#>  [29,] 0.455503619 0.33626992
#>  [30,] 0.600933809 0.64549187
#>  [31,] 0.429181597 0.07756247
#>  [32,] 0.891671680 0.57587259
#>  [33,] 0.551179583 0.42633543
#>  [34,] 0.319364664 0.92745331
#>  [35,] 0.158926877 0.56442975
#>  [36,] 0.000000000 0.77575771
#>  [37,] 0.830078009 0.00832426
#>  [38,] 0.457780416 0.30395176
#>  [39,] 0.511248648 0.58927462
#>  [40,] 0.331518855 0.63010726
#>  [41,] 0.665094146 0.92758251
#>  [42,] 0.379801397 0.26304469
#>  [43,] 0.570689686 0.87103967
#>  [44,] 0.440802661 0.66171352
#>  [45,] 0.240357038 1.00000000
#>  [46,] 0.042431837 0.01650487
#>  [47,] 0.051338303 0.09704640
#>  [48,] 0.648421934 0.10472636
#>  [49,] 0.859980528 0.21244083
#>  [50,] 0.405913711 0.87201231
#>  [51,] 0.674265294 0.55499947
#>  [52,] 0.568239211 0.22790339
#>  [53,] 0.941993173 0.94921684
#>  [54,] 0.143674555 0.33218567
#>  [55,] 0.626505153 0.35913803
#>  [56,] 0.984325183 0.54753006
#>  [57,] 0.945052348 0.86943517
#>  [58,] 0.938860116 0.37429399
#>  [59,] 0.163880024 0.81531479
#>  [60,] 0.016906306 0.81584939
#>  [61,] 0.655550400 0.15435393
#>  [62,] 0.004990548 0.21004585
#>  [63,] 0.565504476 0.68829859
#>  [64,] 0.730674334 0.33108805
#>  [65,] 0.350738059 0.73668302
#>  [66,] 0.064687161 0.51903636
#>  [67,] 0.164802363 0.98364114
#>  [68,] 0.901345061 0.91073078
#>  [69,] 0.760061029 0.12228873
#>  [70,] 0.528794122 0.50917568
#>  [71,] 0.620659969 0.68149268
#>  [72,] 0.507439193 0.05382380
#>  [73,] 0.358015069 0.60447089
#>  [74,] 0.139257838 0.27630199
#>  [75,] 0.463449311 0.92122834
#>  [76,] 0.110336345 0.33461694
#>  [77,] 0.373287922 0.36956137
#>  [78,] 0.795554876 0.87804735
#>  [79,] 0.478250328 0.92981364
#>  [80,] 0.749206139 0.85235847
#>  [81,] 0.931806982 0.06392915
#>  [82,] 0.619497870 0.41462239
#>  [83,] 0.774688881 0.82321841
#>  [84,] 0.196093551 0.51396227
#>  [85,] 0.980812802 0.00000000
#>  [86,] 1.000000000 0.97250373
#>  [87,] 0.590170156 0.82743770
#>  [88,] 0.932702891 0.81776290
#>  [89,] 0.464730460 0.95580398
#>  [90,] 0.750211405 0.68281735
#>  [91,] 0.331096759 0.25110430
#>  [92,] 0.093664681 0.66284925
#>  [93,] 0.544106450 0.58969567
#>  [94,] 0.882722195 0.86833393
#>  [95,] 0.141701822 0.34328150
#>  [96,] 0.388504122 0.45311624
#>  [97,] 0.095099748 0.21096931
#>  [98,] 0.249517059 0.04126542
#>  [99,] 0.318085050 0.83410145
#> [100,] 0.937106278 0.79827905
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,34,99), [4,8,2]
#> 2: (50,34,1), [1,3,10]
#> 3: (1,65,50), [27,2,8]
#> 4: (17,99,34), [1,5,12]
#> 5: (14,17,34), [4,14,7]
#> 6: (89,50,75), [22,16,10]
#> 7: (45,17,14), [5,18,20]
#> 8: (1,99,65), [9,3,1]
#> 9: (99,59,65), [25,8,12]
#> 10: (89,34,50), [2,6,15]
#> 11: (87,50,44), [27,37,33]
#> 12: (99,17,59), [23,9,4]
#> 13: (10,17,67), [20,0,23]
#> 14: (21,14,34), [5,15,18]
#> 15: (21,34,89), [10,19,14]
#> 16: (79,89,75), [6,17,21]
#> 17: (75,2,79), [21,16,22]
#> 18: (14,21,45), [129,7,14]
#> 19: (41,21,89), [15,45,101]
#> 20: (17,45,67), [0,13,7]
#> 21: (89,79,2), [17,45,16]
#> 22: (2,75,50), [6,33,17]
#> 23: (10,59,17), [12,13,30]
#> 24: (65,40,44), [31,27,25]
#> 25: (40,65,59), [9,35,24]
#> 26: (92,59,7), [34,41,35]
#> 27: (44,50,65), [3,24,11]
#> 28: (2,43,41), [53,45,29]
#> 29: (2,87,43), [46,28,33]
#> 30: (59,10,60), [0,34,23]
#> 31: (73,44,40), [24,44,43]
#> 32: (92,6,35), [67,39,36]
#> 33: (2,50,87), [11,29,22]
#> 34: (59,60,7), [40,26,30]
#> 35: (92,40,59), [25,26,39]
#> 36: (36,6,92), [32,41,68]
#> 37: (63,87,44), [11,38,48]
#> 38: (44,39,63), [50,37,43]
#> 39: (92,35,40), [42,35,32]
#> 40: (7,60,36), [0,41,34]
#> 41: (36,92,7), [26,40,36]
#> 42: (40,35,84), [55,51,39]
#> 43: (44,73,39), [58,38,31]
#> 44: (40,9,73), [54,31,51]
#> 45: (41,89,2), [21,28,19]
#> 46: (5,43,87), [29,64,53]
#> 47: (23,87,71), [48,59,56]
#> 48: (87,63,71), [52,47,37]
#> 49: (30,63,93), [50,81,52]
#> 50: (63,39,93), [61,49,38]
#> 51: (40,84,9), [65,44,42]
#> 52: (30,71,63), [48,49,76]
#> 53: (5,41,43), [28,46,77]
#> 54: (73,9,96), [75,58,44]
#> 55: (35,11,84), [65,42,66]
#> 56: (80,87,23), [47,69,64]
#> 57: (51,90,71), [59,76,133]
#> 58: (73,96,39), [62,43,54]
#> 59: (90,23,71), [47,57,71]
#> 60: (51,93,70), [61,88,81]
#> 61: (70,93,39), [50,62,60]
#> 62: (96,70,39), [61,58,73]
#> 63: (5,80,78), [78,77,64]
#> 64: (80,5,87), [46,56,63]
#> 65: (84,11,9), [74,51,55]
#> 66: (35,66,11), [103,55,67]
#> 67: (35,6,66), [68,66,32]
#> 68: (66,6,36), [36,116,67]
#> 69: (3,80,23), [56,70,78]
#> 70: (23,83,3), [79,69,71]
#> 71: (90,83,23), [70,59,85]
#> 72: (33,70,27), [73,83,89]
#> 73: (27,70,96), [62,84,72]
#> 74: (9,11,95), [102,93,65]
#> 75: (9,77,96), [84,54,93]
#> 76: (51,71,30), [52,81,57]
#> 77: (5,78,41), [95,53,63]
#> 78: (78,80,3), [69,79,63]
#> 79: (78,3,83), [70,80,78]
#> 80: (83,18,78), [97,79,85]
#> 81: (51,30,93), [49,60,76]
#> 82: (70,82,4), [111,88,89]
#> 83: (27,29,33), [105,72,90]
#> 84: (77,27,96), [73,75,90]
#> 85: (83,90,18), [109,80,71]
#> 86: (26,19,90), [109,87,113]
#> 87: (32,26,90), [86,133,132]
#> 88: (4,51,70), [60,82,114]
#> 89: (70,33,82), [112,82,72]
#> 90: (27,77,29), [99,83,84]
#> 91: (55,29,38), [106,142,105]
#> 92: (54,77,95), [93,100,118]
#> 93: (9,95,77), [92,75,74]
#> 94: (11,16,76), [107,102,103]
#> 95: (78,13,41), [130,77,104]
#> 96: (68,78,94), [97,124,104]
#> 97: (94,78,18), [80,98,96]
#> 98: (18,88,94), [108,97,122]
#> 99: (29,77,42), [121,106,90]
#> 100: (95,76,54), [115,92,102]
#> 101: (28,21,41), [19,130,129]
#> 102: (76,95,11), [74,94,100]
#> 103: (11,66,16), [116,94,66]
#> 104: (68,13,78), [95,96,127]
#> 105: (55,33,29), [83,91,112]
#> 106: (29,42,38), [117,91,99]
#> 107: (74,76,16), [94,146,115]
#> 108: (57,94,88), [98,125,124]
#> 109: (19,18,90), [85,86,119]
#> 110: (64,4,55), [111,155,137]
#> 111: (55,4,82), [82,112,110]
#> 112: (55,82,33), [89,105,111]
#> 113: (100,19,26), [86,153,119]
#> 114: (51,4,32), [137,133,88]
#> 115: (74,54,76), [100,107,120]
#> 116: (66,36,16), [150,103,68]
#> 117: (38,42,20), [135,134,106]
#> 118: (22,77,54), [92,120,128]
#> 119: (18,19,100), [113,122,109]
#> 120: (22,54,74), [115,144,118]
#> 121: (77,91,42), [143,99,128]
#> 122: (18,100,88), [125,98,119]
#> 123: (68,57,53), [139,127,124]
#> 124: (57,68,94), [96,108,123]
#> 125: (57,88,100), [122,140,108]
#> 126: (13,53,28), [131,130,127]
#> 127: (68,53,13), [126,104,123]
#> 128: (22,91,77), [121,118,136]
#> 129: (28,45,21), [18,101,0]
#> 130: (28,41,13), [95,126,101]
#> 131: (53,86,28), [0,126,139]
#> 132: (56,26,32), [87,151,153]
#> 133: (32,90,51), [57,114,87]
#> 134: (52,38,20), [117,159,142]
#> 135: (31,20,42), [117,143,166]
#> 136: (98,91,22), [128,147,164]
#> 137: (64,32,4), [114,110,154]
#> 138: (12,55,52), [142,156,155]
#> 139: (53,57,86), [140,131,123]
#> 140: (86,57,100), [125,141,139]
#> 141: (100,56,86), [185,140,153]
#> 142: (52,55,38), [91,134,138]
#> 143: (31,42,91), [121,164,135]
#> 144: (74,15,22), [147,120,145]
#> 145: (74,97,15), [163,144,146]
#> 146: (97,74,16), [107,149,145]
#> 147: (98,22,15), [144,162,136]
#> 148: (47,97,62), [149,174,163]
#> 149: (62,97,16), [146,150,148]
#> 150: (62,16,36), [116,0,149]
#> 151: (58,56,32), [132,154,183]
#> 152: (64,12,8), [158,169,155]
#> 153: (56,100,26), [113,132,141]
#> 154: (58,32,64), [137,160,151]
#> 155: (64,55,12), [138,152,110]
#> 156: (52,61,12), [165,138,161]
#> 157: (61,48,69), [171,165,161]
#> 158: (69,8,12), [152,165,170]
#> 159: (20,72,52), [167,134,166]
#> 160: (64,49,58), [180,154,169]
#> 161: (52,48,61), [157,156,167]
#> 162: (98,15,47), [163,173,147]
#> 163: (47,15,97), [145,148,162]
#> 164: (91,98,31), [168,143,136]
#> 165: (69,12,61), [156,157,158]
#> 166: (20,31,72), [168,159,135]
#> 167: (52,72,48), [177,161,159]
#> 168: (31,98,72), [178,166,164]
#> 169: (64,8,49), [170,160,152]
#> 170: (69,49,8), [169,158,175]
#> 171: (48,25,69), [176,157,177]
#> 172: (37,98,46), [173,188,178]
#> 173: (46,98,47), [162,174,172]
#> 174: (46,47,62), [148,0,173]
#> 175: (69,24,49), [179,170,176]
#> 176: (69,25,24), [181,175,171]
#> 177: (48,72,25), [182,171,167]
#> 178: (72,98,37), [172,182,168]
#> 179: (49,24,81), [184,180,175]
#> 180: (49,81,58), [186,160,179]
#> 181: (37,24,25), [176,182,187]
#> 182: (72,37,25), [181,177,178]
#> 183: (58,85,56), [185,151,186]
#> 184: (24,85,81), [186,179,187]
#> 185: (85,86,56), [141,183,0]
#> 186: (58,81,85), [184,183,180]
#> 187: (24,37,85), [188,184,181]
#> 188: (85,37,46), [172,0,187]
#> boundary nodes:  67 10 60 36 62 46 85 86 28 45 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.09336063 0.08456140 0.14614503 0.12590736 0.07571369 0.12867029 0.10274517 
#>                                                                              
#> 0.20272859 0.10278347 0.21317322 0.07627395 0.10903503 0.08112831 0.14810461 
#>                                                                              
#> 0.03459937 0.08286782 0.20961226 0.04007747 0.07730538 0.11376250 0.11607034 
#>                                                                              
#> 0.11004224 0.04775584 0.08747746 0.10654788 0.18957145 0.10257235 0.24980914 
#>                                                                              
#> 0.02037699 0.22310346 0.11291343 0.18466599 0.04351052 0.07344613 0.06267638 
#>                                                                              
#> 0.15398525 0.13775174 0.10843481 0.14909586 0.05557381 0.10089803 0.03286050 
#>                                                                              
#> 0.12743673 0.07443913 0.10286029 0.13706253 0.12955821 0.18331754 0.06700781 
#>                                                                              
#> 0.08196302 0.08199807 0.05299174 0.12023298 0.19304903 0.10238081 0.12647233 
#>                                                                              
#> 0.15860053 0.15143832 0.02192634 0.14312702 0.09057868 0.15107596 0.09564329 
#>                                                                              
#> 0.08492908 0.04105095 0.13954832 0.01693981 0.08395925 0.07964022 0.05636080 
#>                                                                              
#> 0.13139896 0.08716599 0.14520462 0.15740124 0.15668162 0.15251781 0.06931510 
#>                                                                              
#> 0.08870029 0.03239825 0.23307242 0.07675263 0.22230932 0.08770685 0.07110203 
#>                                                                              
#> 0.10671565 0.03342674 0.22496785 0.19243178 0.11575399 0.08805746 0.09738680 
#>                                                                              
#> 0.01082133 0.26048255 0.44793091 0.10157515 0.23263542 0.07883881 0.05605795 
#>                                                                              
#> 0.05014693 0.01997524 0.07984030 0.04630664 0.04443487 0.05572153 0.12574646 
#>                                                                              
#> 0.05290937 0.14219187 0.14867826 0.06190535 0.07106090 0.14358779 0.11680057 
#>                                                                              
#> 0.07159842 0.42526264 0.17752059 0.10140363 0.09385204 0.11795352 0.09011393 
#>                                                                              
#> 0.13309888 0.09689154 0.06981527 0.08758668 0.29298399 0.14609980 0.07634725 
#>                                                                              
#> 0.11301205 0.18439183 0.18005121 0.05013692 0.15263239 0.18329436 0.08177882 
#>                                                                              
#> 0.14989064 0.25822769 0.13482037 0.11037923 0.20856024 0.20587942 0.12213513 
#>                                                                              
#> 0.15330044 0.09417587 0.30242494 0.58149474 0.03821255 0.31044497 0.02627190 
#>                                                                              
#> 0.54754133 0.08055127 0.15096447 0.78768865 0.02958632 0.06610356 0.03010093 
#>                                                                              
#> 0.11028579 0.10237409 0.15534519 0.13622953 0.14603387 0.01711074 0.20234144 
#>                                                                              
#> 0.14616011 0.12905400 0.11718387 0.12014973 0.02740267 0.14697469 0.03686832 
#>                                                                              
#> 0.11809128 0.07388081 0.16785689 0.14670172 0.12750415 0.24009716 0.08982666 
#>                                                                              
#> 0.17840872 0.10104470 0.15776150 0.05556666 0.11790028 0.15439499 0.13738291 
#>                                                                              
#> 0.11369089 0.16097667 0.13061343 0.10460251 0.26474569 0.14830227 0.08357505 
#>                                                                              
#> 0.07775056 0.21791841 0.11647481 0.03200934 0.05039917 0.05866517 0.13470394 
#>                                                                              
#> 0.14821959 0.06122796 0.17733638 0.13102478 0.10237622 0.01126983 0.11789884 
#>                                                                              
#> 0.10450083 0.08285012 0.10532218 0.29332738 0.14316756 0.11072385 0.17252384 
#>                                                                              
#> 0.05312752 0.10787965 0.05592509 0.10094272 0.21840610 0.06509290 0.14952558 
#>                                                                              
#> 0.10486382 0.12718210 0.10151579 0.05597719 0.06233988 0.07766121 0.04200544 
#>                                                                              
#> 0.17184651 0.50124751 0.09014383 0.19194285 0.19854684 0.13528749 0.11624630 
#>                                                                              
#> 0.06250657 0.18522921 0.13410631 0.19934242 0.10927940 0.15045499 0.12248853 
#>                                                                              
#> 0.56573388 0.20702821 0.06169080 0.18853531 0.21262190 0.11570258 0.10931902 
#>                                                                              
#> 0.12755718 0.11416114 0.12203890 0.19428611 0.14697567 0.17549147 0.14918563 
#>                                                                              
#> 0.08103248 0.19712930 0.13457826 0.14616817 0.17954472 0.32583127 0.16496896 
#>                                                                              
#> 0.17910279 0.37663778 0.10750665 0.93852610 0.12506157 0.12472796 0.18073252 
#>                                                                              
#> 0.02929649 0.09289443 0.11303780 0.10829483 0.10065040 0.14130848 0.08165415 
#>                                                                              
#> 0.14251879 0.08581151 0.12129720 0.10654739 0.03254025 0.10114181 0.06013031 
#>                                                                              
#> 0.71078075 0.17607649 0.22513907 0.25515619 0.07885617 0.08373703 0.97269299 
#> 
#> $alpha
#> [1] 0.2638547
#> 
#> attr(,"class")
#> [1] "scree"