Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1         y1           x2         y2          mx1
#>     8   28 0.7561173224 0.82602123 0.6415949106 0.89842217  0.675832937
#>    28   86 0.6415949106 0.89842217 0.5933671197 0.96775521  0.576366828
#>    86   43 0.5933671197 0.96775521 0.5412672843 0.80106434  0.509734330
#>    43   27 0.5412672843 0.80106434 0.5462584992 0.73723729  0.604578422
#>    27   23 0.5462584992 0.73723729 0.5894986266 0.68459883  0.629712892
#>    23    2 0.5894986266 0.68459883 0.7205058203 0.67462929  0.661131864
#>     2    8 0.7205058203 0.67462929 0.7561173224 0.82602123  0.669710126
#>    73   24 0.6350347309 0.62636453 0.7517913868 0.62055178  0.693382738
#>    24   18 0.7517913868 0.62055178 0.7564913888 0.69762731  0.755845207
#>    18    8 0.7564913888 0.69762731 0.7561173224 0.82602123  0.690245059
#>    23   73 0.5894986266 0.68459883 0.6350347309 0.62636453  0.655753471
#>    53   92 0.0180922888 0.00000000 0.0454993380 0.14235142 -1.030229814
#>    92   60 0.0454993380 0.14235142 0.0387269613 0.47007565 -0.919744069
#>    60   90 0.0387269613 0.47007565 0.1333137421 0.67109063  0.100983779
#>    90    5 0.1333137421 0.67109063 0.0737224347 0.72684747  0.061378033
#>     5   47 0.0737224347 0.72684747 0.0263164010 0.79908991 -0.028098070
#>    47   79 0.0263164010 0.79908991 0.0030942567 0.82701368 -0.095684446
#>    79   59 0.0030942567 0.82701368 0.0003320027 0.94711724 -4.465265278
#>    86   38 0.5933671197 0.96775521 0.4241587475 0.99252130  0.499546553
#>    38   69 0.4241587475 0.99252130 0.3172661533 0.93635745  0.381166960
#>    69   19 0.3172661533 0.93635745 0.3896694936 0.82918869  0.329832830
#>    19   43 0.3896694936 0.82918869 0.5412672843 0.80106434  0.470620411
#>    94   47 0.1334025267 0.84472883 0.0263164010 0.79908991  0.092967977
#>    90   29 0.1333137421 0.67109063 0.1883382096 0.69531280  0.137770147
#>    29   94 0.1883382096 0.69531280 0.1334025267 0.84472883  0.145713671
#>    87   56 0.3667933280 0.62849070 0.3957822850 0.53545901  0.434407968
#>    56   17 0.3957822850 0.53545901 0.4774403959 0.45849327  0.444494823
#>    17   75 0.4774403959 0.45849327 0.5345130893 0.47950197  0.494332856
#>    75   96 0.5345130893 0.47950197 0.5602107084 0.48206384  0.541857561
#>    96   68 0.5602107084 0.48206384 0.6210228704 0.54613833  0.553301440
#>    68   42 0.6210228704 0.54613833 0.5566768822 0.63045595  0.553219819
#>    42   87 0.5566768822 0.63045595 0.3667933280 0.62849070  0.461653126
#>    78   97 0.8894146739 0.65868827 0.9033413516 0.70073476  0.914078033
#>    97   63 0.9033413516 0.70073476 0.9008808487 0.70688075  0.885714528
#>    63   58 0.9008808487 0.70688075 0.8191328940 0.72833174  0.858592553
#>    58   78 0.8191328940 0.72833174 0.8894146739 0.65868827  0.817189794
#>    48   36 0.8203543144 0.76272581 0.7781499711 0.88240873  0.824943009
#>    36   37 0.7781499711 0.88240873 0.7605967434 0.86728885  0.797299779
#>    37   98 0.7605967434 0.86728885 0.7518715944 0.85532806  0.779462078
#>    98   28 0.7518715944 0.85532806 0.6415949106 0.89842217  0.678301552
#>    18   48 0.7564913888 0.69762731 0.8203543144 0.76272581  0.756163265
#>    85   91 0.2301740789 0.96351407 0.1513435207 0.93938651  0.191588477
#>    91   70 0.1513435207 0.93938651 0.2193994191 0.87221021  0.177866316
#>    70   69 0.2193994191 0.87221021 0.3172661533 0.93635745  0.275213733
#>    69   21 0.3172661533 0.93635745 0.2871507781 0.95567091  0.273866872
#>    21   85 0.2871507781 0.95567091 0.2301740789 0.96351407  0.255419926
#>    54   29 0.2994946916 0.76188429 0.1883382096 0.69531280  0.258932842
#>    29   62 0.1883382096 0.69531280 0.3256012725 0.56383665  0.257654328
#>    62   72 0.3256012725 0.56383665 0.3537011038 0.61717375  0.264497668
#>    72   87 0.3537011038 0.61717375 0.3667933280 0.62849070  0.334401868
#>    87   33 0.3667933280 0.62849070 0.3819671007 0.79670376  0.428528364
#>    33   54 0.3819671007 0.79670376 0.2994946916 0.76188429  0.355226785
#>    60   57 0.0387269613 0.47007565 0.0559119431 0.25189104 -0.290069902
#>    57   50 0.0559119431 0.25189104 0.0912461441 0.29629604  0.111179431
#>    50   39 0.0912461441 0.29629604 0.0527925782 0.45004688 -0.005419819
#>    39   60 0.0527925782 0.45004688 0.0387269613 0.47007565 -0.109572715
#>    57   92 0.0559119431 0.25189104 0.0454993380 0.14235142  0.111751121
#>    92   76 0.0454993380 0.14235142 0.1277764854 0.04489962  0.091757666
#>    76   32 0.1277764854 0.04489962 0.1637980395 0.08570702  0.118814702
#>    32   66 0.1637980395 0.08570702 0.2349388380 0.11581718  0.193413959
#>    66   50 0.2349388380 0.11581718 0.0912461441 0.29629604  0.198757699
#>    82   84 0.5609531412 0.34681124 0.5472187183 0.33826217  0.583011504
#>    84   83 0.5472187183 0.33826217 0.5039174253 0.17110982  0.535241871
#>    83   14 0.5039174253 0.17110982 0.6147469623 0.12667896  0.577433715
#>    14  100 0.6147469623 0.12667896 0.7489918901 0.24687103  0.680522048
#>   100   31 0.7489918901 0.24687103 0.6381457667 0.28920434  0.685024856
#>    31   82 0.6381457667 0.28920434 0.5609531412 0.34681124  0.583343425
#>    83   30 0.5039174253 0.17110982 0.5964702347 0.08136506  0.550288167
#>    30   55 0.5964702347 0.08136506 0.8865816942 0.15180529  0.759663162
#>    55  100 0.8865816942 0.15180529 0.7489918901 0.24687103  0.753684958
#>    56   62 0.3957822850 0.53545901 0.3256012725 0.56383665  0.344006275
#>    62   46 0.3256012725 0.56383665 0.3088384173 0.43634678  0.327348107
#>    46   51 0.3088384173 0.43634678 0.4547920658 0.39546084  0.372239935
#>    51   17 0.4547920658 0.39546084 0.4774403959 0.45849327  0.428745038
#>    46   71 0.3088384173 0.43634678 0.2784363936 0.45510486  0.273823827
#>    71   95 0.2784363936 0.45510486 0.1699122543 0.42760604  0.227322422
#>    95   52 0.1699122543 0.42760604 0.1887323849 0.37124975  0.217129298
#>    52   66 0.1887323849 0.37124975 0.2349388380 0.11581718  0.233183319
#>    66   99 0.2349388380 0.11581718 0.4344154721 0.19976869  0.298861620
#>    99   26 0.4344154721 0.19976869 0.4472611092 0.28493711  0.316667327
#>    26   46 0.4472611092 0.28493711 0.3088384173 0.43634678  0.353304889
#>    51   82 0.4547920658 0.39546084 0.5609531412 0.34681124  0.523941040
#>    82   75 0.5609531412 0.34681124 0.5345130893 0.47950197  0.527980852
#>    24   58 0.7517913868 0.62055178 0.8191328940 0.72833174  0.816078175
#>    58   48 0.8191328940 0.72833174 0.8203543144 0.76272581  0.771009461
#>    69   54 0.3172661533 0.93635745 0.2994946916 0.76188429  0.304275899
#>    33   43 0.3819671007 0.79670376 0.5412672843 0.80106434  0.461724928
#>    49   65 0.9126161216 0.84416206 0.9483415576 0.74110308  1.195614908
#>    65   41 0.9483415576 0.74110308 0.9428221056 0.69384107  0.980248136
#>    41   22 0.9428221056 0.69384107 0.9502122337 0.68135201  0.978016095
#>    22   78 0.9502122337 0.68135201 0.8894146739 0.65868827  0.962169674
#>    78   25 0.8894146739 0.65868827 0.9840677985 0.36641406  0.967500502
#>    25   88 0.9840677985 0.36641406 0.9855332303 0.08630157  6.194117272
#>    80   85 0.2928683309 1.00000000 0.2301740789 0.96351407  0.261676717
#>    69   80 0.3172661533 0.93635745 0.2928683309 1.00000000  0.320124909
#>    41   78 0.9428221056 0.69384107 0.8894146739 0.65868827  0.919271338
#>    27   42 0.5462584992 0.73723729 0.5566768822 0.63045595  0.532693164
#>    42   67 0.5566768822 0.63045595 0.5935759512 0.66513076  0.556569004
#>    67   73 0.5935759512 0.66513076 0.6350347309 0.62636453  0.653706321
#>    73   68 0.6350347309 0.62636453 0.6210228704 0.54613833  0.691013055
#>    68   44 0.6210228704 0.54613833 0.7698363334 0.46277471  0.713223321
#>    44   78 0.7698363334 0.46277471 0.8894146739 0.65868827  0.845077089
#>    44   45 0.7698363334 0.46277471 0.8046203971 0.38603997  0.905475878
#>    45   64 0.8046203971 0.38603997 0.8918356175 0.34207124  0.905420249
#>    64   74 0.8918356175 0.34207124 0.9043695325 0.21767850  0.956038841
#>    74   88 0.9043695325 0.21767850 0.9855332303 0.08630157  1.066015736
#>    84   51 0.5472187183 0.33826217 0.4547920658 0.39546084  0.483150554
#>    99   83 0.4344154721 0.19976869 0.5039174253 0.17110982  0.489601840
#>    98   61 0.7518715944 0.85532806 0.6872243966 0.91389809  0.684745998
#>    61   77 0.6872243966 0.91389809 0.6287146228 0.95985265  0.655191674
#>    77   86 0.6287146228 0.95985265 0.5933671197 0.96775521  0.601723425
#>    90   62 0.1333137421 0.67109063 0.3256012725 0.56383665  0.207272816
#>    54   70 0.2994946916 0.76188429 0.2193994191 0.87221021  0.211703290
#>    70   94 0.2193994191 0.87221021 0.1334025267 0.84472883  0.200061193
#>    88   55 0.9855332303 0.08630157 0.8865816942 0.15180529  0.793291270
#>    83   93 0.5039174253 0.17110982 0.2118906512 0.06023412  0.409127848
#>    93   53 0.2118906512 0.06023412 0.0180922888 0.00000000  0.468998034
#>    53   88 0.0180922888 0.00000000 0.9855332303 0.08630157  0.839556948
#>   100   34 0.7489918901 0.24687103 0.7470900156 0.27723734  0.689892793
#>    34   40 0.7470900156 0.27723734 0.7206605761 0.42298787  0.699276470
#>    40   82 0.7206605761 0.42298787 0.5609531412 0.34681124  0.644184478
#>    76   93 0.1277764854 0.04489962 0.2118906512 0.06023412  0.171678133
#>    93   66 0.2118906512 0.06023412 0.2349388380 0.11581718  0.200788010
#>   100   74 0.7489918901 0.24687103 0.9043695325 0.21767850  0.833278709
#>    45   40 0.8046203971 0.38603997 0.7206605761 0.42298787  0.739118709
#>    89   61 0.6458990470 0.97985955 0.6872243966 0.91389809  0.725281217
#>    61   81 0.6872243966 0.91389809 0.8015939189 0.92300937  0.742151784
#>    81   49 0.8015939189 0.92300937 0.9126161216 0.84416206  0.898547853
#>    48   49 0.8203543144 0.76272581 0.9126161216 0.84416206  0.837653108
#>    61   37 0.6872243966 0.91389809 0.7605967434 0.86728885  0.737450926
#>    86   89 0.5933671197 0.96775521 0.6458990470 0.97985955  0.551264670
#>    50   95 0.0912461441 0.29629604 0.1699122543 0.42760604  0.100119094
#>    95   60 0.1699122543 0.42760604 0.0387269613 0.47007565  0.126252131
#>    96   82 0.5602107084 0.48206384 0.5609531412 0.34681124  0.626545239
#>    65   63 0.9483415576 0.74110308 0.9008808487 0.70688075  0.927816433
#>    62   71 0.3256012725 0.56383665 0.2784363936 0.45510486  0.326433737
#>    94   79 0.1334025267 0.84472883 0.0030942567 0.82701368  0.065596948
#>    94   91 0.1334025267 0.84472883 0.1513435207 0.93938651  0.073881543
#>    91   80 0.1513435207 0.93938651 0.2928683309 1.00000000  0.092663513
#>    95   90 0.1699122543 0.42760604 0.1333137421 0.67109063  0.136102918
#>    89   77 0.6458990470 0.97985955 0.6287146228 0.95985265  0.665114017
#>    93   99 0.2118906512 0.06023412 0.4344154721 0.19976869  0.393768491
#>     1   28 0.5972842675 0.84240939 0.6415949106 0.89842217  0.675832937
#>     1   86 0.5972842675 0.84240939 0.5933671197 0.96775521  0.576366828
#>     1   43 0.5972842675 0.84240939 0.5412672843 0.80106434  0.509734330
#>     1   27 0.5972842675 0.84240939 0.5462584992 0.73723729  0.604578422
#>     1   23 0.5972842675 0.84240939 0.5894986266 0.68459883  0.629712892
#>     1    2 0.5972842675 0.84240939 0.7205058203 0.67462929  0.661131864
#>     1    8 0.5972842675 0.84240939 0.7561173224 0.82602123  0.669710126
#>     2   24 0.7205058203 0.67462929 0.7517913868 0.62055178  0.693382738
#>     2   18 0.7205058203 0.67462929 0.7564913888 0.69762731  0.755845207
#>     2   73 0.7205058203 0.67462929 0.6350347309 0.62636453  0.655753471
#>     3   92 0.0000000000 0.61267643 0.0454993380 0.14235142 -1.030229814
#>     3   60 0.0000000000 0.61267643 0.0387269613 0.47007565 -0.919744069
#>     3   90 0.0000000000 0.61267643 0.1333137421 0.67109063  0.100983779
#>     3    5 0.0000000000 0.61267643 0.0737224347 0.72684747  0.061378033
#>     3   47 0.0000000000 0.61267643 0.0263164010 0.79908991 -0.028098070
#>     3   79 0.0000000000 0.61267643 0.0030942567 0.82701368 -0.095684446
#>     3   59 0.0000000000 0.61267643 0.0003320027 0.94711724 -4.465265278
#>     4   38 0.3987328284 0.88256028 0.4241587475 0.99252130  0.499546553
#>     4   69 0.3987328284 0.88256028 0.3172661533 0.93635745  0.381166960
#>     4   19 0.3987328284 0.88256028 0.3896694936 0.82918869  0.329832830
#>     4   43 0.3987328284 0.88256028 0.5412672843 0.80106434  0.470620411
#>     4   86 0.3987328284 0.88256028 0.5933671197 0.96775521  0.505417586
#>     5   29 0.0737224347 0.72684747 0.1883382096 0.69531280  0.137770147
#>     5   94 0.0737224347 0.72684747 0.1334025267 0.84472883  0.145713671
#>     6   56 0.4856343048 0.54518488 0.3957822850 0.53545901  0.434407968
#>     6   17 0.4856343048 0.54518488 0.4774403959 0.45849327  0.444494823
#>     6   75 0.4856343048 0.54518488 0.5345130893 0.47950197  0.494332856
#>     6   96 0.4856343048 0.54518488 0.5602107084 0.48206384  0.541857561
#>     6   68 0.4856343048 0.54518488 0.6210228704 0.54613833  0.553301440
#>     6   42 0.4856343048 0.54518488 0.5566768822 0.63045595  0.553219819
#>     6   87 0.4856343048 0.54518488 0.3667933280 0.62849070  0.461653126
#>     7   97 0.8856405630 0.67927441 0.9033413516 0.70073476  0.914078033
#>     7   63 0.8856405630 0.67927441 0.9008808487 0.70688075  0.885714528
#>     7   58 0.8856405630 0.67927441 0.8191328940 0.72833174  0.858592553
#>     7   78 0.8856405630 0.67927441 0.8894146739 0.65868827  0.817189794
#>     8   36 0.7561173224 0.82602123 0.7781499711 0.88240873  0.824943009
#>     8   37 0.7561173224 0.82602123 0.7605967434 0.86728885  0.797299779
#>     8   98 0.7561173224 0.82602123 0.7518715944 0.85532806  0.779462078
#>     8   48 0.7561173224 0.82602123 0.8203543144 0.76272581  0.756163265
#>     9   91 0.2215783580 0.92031872 0.1513435207 0.93938651  0.191588477
#>     9   70 0.2215783580 0.92031872 0.2193994191 0.87221021  0.177866316
#>     9   69 0.2215783580 0.92031872 0.3172661533 0.93635745  0.275213733
#>     9   21 0.2215783580 0.92031872 0.2871507781 0.95567091  0.273866872
#>     9   85 0.2215783580 0.92031872 0.2301740789 0.96351407  0.255419926
#>    10   29 0.3291978092 0.69285374 0.1883382096 0.69531280  0.258932842
#>    10   62 0.3291978092 0.69285374 0.3256012725 0.56383665  0.257654328
#>    10   72 0.3291978092 0.69285374 0.3537011038 0.61717375  0.264497668
#>    10   87 0.3291978092 0.69285374 0.3667933280 0.62849070  0.334401868
#>    10   33 0.3291978092 0.69285374 0.3819671007 0.79670376  0.428528364
#>    10   54 0.3291978092 0.69285374 0.2994946916 0.76188429  0.355226785
#>    11   57 0.0579291596 0.26086098 0.0559119431 0.25189104 -0.290069902
#>    11   50 0.0579291596 0.26086098 0.0912461441 0.29629604  0.111179431
#>    11   39 0.0579291596 0.26086098 0.0527925782 0.45004688 -0.005419819
#>    11   60 0.0579291596 0.26086098 0.0387269613 0.47007565 -0.109572715
#>    12   92 0.1521424482 0.11951586 0.0454993380 0.14235142  0.111751121
#>    12   76 0.1521424482 0.11951586 0.1277764854 0.04489962  0.091757666
#>    12   32 0.1521424482 0.11951586 0.1637980395 0.08570702  0.118814702
#>    12   66 0.1521424482 0.11951586 0.2349388380 0.11581718  0.193413959
#>    12   50 0.1521424482 0.11951586 0.0912461441 0.29629604  0.198757699
#>    12   57 0.1521424482 0.11951586 0.0559119431 0.25189104  0.146164591
#>    13   84 0.6220286070 0.25683321 0.5472187183 0.33826217  0.583011504
#>    13   83 0.6220286070 0.25683321 0.5039174253 0.17110982  0.535241871
#>    13   14 0.6220286070 0.25683321 0.6147469623 0.12667896  0.577433715
#>    13  100 0.6220286070 0.25683321 0.7489918901 0.24687103  0.680522048
#>    13   31 0.6220286070 0.25683321 0.6381457667 0.28920434  0.685024856
#>    13   82 0.6220286070 0.25683321 0.5609531412 0.34681124  0.583343425
#>    14   30 0.6147469623 0.12667896 0.5964702347 0.08136506  0.550288167
#>    14   55 0.6147469623 0.12667896 0.8865816942 0.15180529  0.759663162
#>    15   62 0.3828484546 0.46473455 0.3256012725 0.56383665  0.344006275
#>    15   46 0.3828484546 0.46473455 0.3088384173 0.43634678  0.327348107
#>    15   51 0.3828484546 0.46473455 0.4547920658 0.39546084  0.372239935
#>    15   17 0.3828484546 0.46473455 0.4774403959 0.45849327  0.428745038
#>    15   56 0.3828484546 0.46473455 0.3957822850 0.53545901  0.432166508
#>    16   71 0.2528731331 0.37750341 0.2784363936 0.45510486  0.273823827
#>    16   95 0.2528731331 0.37750341 0.1699122543 0.42760604  0.227322422
#>    16   52 0.2528731331 0.37750341 0.1887323849 0.37124975  0.217129298
#>    16   66 0.2528731331 0.37750341 0.2349388380 0.11581718  0.233183319
#>    16   99 0.2528731331 0.37750341 0.4344154721 0.19976869  0.298861620
#>    16   26 0.2528731331 0.37750341 0.4472611092 0.28493711  0.316667327
#>    16   46 0.2528731331 0.37750341 0.3088384173 0.43634678  0.353304889
#>    17   82 0.4774403959 0.45849327 0.5609531412 0.34681124  0.523941040
#>    18   58 0.7564913888 0.69762731 0.8191328940 0.72833174  0.816078175
#>    19   54 0.3896694936 0.82918869 0.2994946916 0.76188429  0.304275899
#>    19   33 0.3896694936 0.82918869 0.3819671007 0.79670376  0.319927022
#>    20   65 1.0000000000 0.67608666 0.9483415576 0.74110308  1.195614908
#>    20   41 1.0000000000 0.67608666 0.9428221056 0.69384107  0.980248136
#>    20   22 1.0000000000 0.67608666 0.9502122337 0.68135201  0.978016095
#>    20   78 1.0000000000 0.67608666 0.8894146739 0.65868827  0.962169674
#>    20   25 1.0000000000 0.67608666 0.9840677985 0.36641406  0.967500502
#>    20   88 1.0000000000 0.67608666 0.9855332303 0.08630157  6.194117272
#>    21   80 0.2871507781 0.95567091 0.2928683309 1.00000000  0.320124909
#>    23   42 0.5894986266 0.68459883 0.5566768822 0.63045595  0.532693164
#>    23   67 0.5894986266 0.68459883 0.5935759512 0.66513076  0.556569004
#>    24   68 0.7517913868 0.62055178 0.6210228704 0.54613833  0.691013055
#>    24   44 0.7517913868 0.62055178 0.7698363334 0.46277471  0.713223321
#>    24   78 0.7517913868 0.62055178 0.8894146739 0.65868827  0.845077089
#>    25   44 0.9840677985 0.36641406 0.7698363334 0.46277471  0.918330954
#>    25   45 0.9840677985 0.36641406 0.8046203971 0.38603997  0.905475878
#>    25   64 0.9840677985 0.36641406 0.8918356175 0.34207124  0.905420249
#>    25   74 0.9840677985 0.36641406 0.9043695325 0.21767850  0.956038841
#>    26   51 0.4472611092 0.28493711 0.4547920658 0.39546084  0.483150554
#>    26   83 0.4472611092 0.28493711 0.5039174253 0.17110982  0.489601840
#>    26   84 0.4472611092 0.28493711 0.5472187183 0.33826217  0.527928187
#>    27   33 0.5462584992 0.73723729 0.3819671007 0.79670376  0.462604815
#>    27   87 0.5462584992 0.73723729 0.3667933280 0.62849070  0.442217075
#>    28   61 0.6415949106 0.89842217 0.6872243966 0.91389809  0.684745998
#>    28   77 0.6415949106 0.89842217 0.6287146228 0.95985265  0.655191674
#>    29   70 0.1883382096 0.69531280 0.2193994191 0.87221021  0.211703290
#>    30   93 0.5964702347 0.08136506 0.2118906512 0.06023412  0.409127848
#>    30   53 0.5964702347 0.08136506 0.0180922888 0.00000000  0.468998034
#>    30   88 0.5964702347 0.08136506 0.9855332303 0.08630157  0.839556948
#>    31   34 0.6381457667 0.28920434 0.7470900156 0.27723734  0.689892793
#>    31   40 0.6381457667 0.28920434 0.7206605761 0.42298787  0.699276470
#>    32   93 0.1637980395 0.08570702 0.2118906512 0.06023412  0.171678133
#>    34   74 0.7470900156 0.27723734 0.9043695325 0.21767850  0.833278709
#>    34   64 0.7470900156 0.27723734 0.8918356175 0.34207124  0.835621424
#>    34   45 0.7470900156 0.27723734 0.8046203971 0.38603997  0.820092908
#>    35   61 0.8029530476 0.96684108 0.6872243966 0.91389809  0.725281217
#>    35   81 0.8029530476 0.96684108 0.8015939189 0.92300937  0.742151784
#>    35   49 0.8029530476 0.96684108 0.9126161216 0.84416206  0.898547853
#>    36   49 0.7781499711 0.88240873 0.9126161216 0.84416206  0.837653108
#>    36   81 0.7781499711 0.88240873 0.8015939189 0.92300937  0.847183593
#>    36   61 0.7781499711 0.88240873 0.6872243966 0.91389809  0.743531769
#>    38   89 0.4241587475 0.99252130 0.6458990470 0.97985955  0.551264670
#>    39   95 0.0527925782 0.45004688 0.1699122543 0.42760604  0.100119094
#>    40   44 0.7206605761 0.42298787 0.7698363334 0.46277471  0.767449325
#>    40   68 0.7206605761 0.42298787 0.6210228704 0.54613833  0.695429569
#>    40   96 0.7206605761 0.42298787 0.5602107084 0.48206384  0.644335392
#>    41   63 0.9428221056 0.69384107 0.9008808487 0.70688075  0.927816433
#>    41   97 0.9428221056 0.69384107 0.9033413516 0.70073476  0.925881804
#>    42   73 0.5566768822 0.63045595 0.6350347309 0.62636453  0.593965143
#>    48   63 0.8203543144 0.76272581 0.9008808487 0.70688075  0.866894472
#>    48   65 0.8203543144 0.76272581 0.9483415576 0.74110308  0.888168929
#>    50   52 0.0912461441 0.29629604 0.1887323849 0.37124975  0.209672078
#>    53   76 0.0180922888 0.00000000 0.1277764854 0.04489962  0.257536890
#>    55   74 0.8865816942 0.15180529 0.9043695325 0.21767850  0.966793375
#>    56   72 0.3957822850 0.53545901 0.3537011038 0.61717375  0.392534978
#>    59   94 0.0003320027 0.94711724 0.1334025267 0.84472883  0.061102947
#>    59   91 0.0003320027 0.94711724 0.1513435207 0.93938651  0.073881543
#>    59   80 0.0003320027 0.94711724 0.2928683309 1.00000000  0.092663513
#>    62   95 0.3256012725 0.56383665 0.1699122543 0.42760604  0.195111115
#>     3   53 0.0000000000 0.61267643 0.0180922888 0.00000000 -1.030229814
#>    20   49 1.0000000000 0.67608666 0.9126161216 0.84416206  1.195614908
#>    35   89 0.8029530476 0.96684108 0.6458990470 0.97985955  0.725281217
#>    38   80 0.4241587475 0.99252130 0.2928683309 1.00000000  0.358071368
#>          my1          mx2         my2 bp1 bp2
#>   0.82580408  0.678301552  0.82970889   0   0
#>   0.90448982  0.601723425  0.92212777   0   0
#>   0.90240751  0.505417586  0.90375672   0   0
#>   0.77390653  0.462604815  0.76280433   0   0
#>   0.76171218  0.532693164  0.68201484   0   0
#>   0.76016211  0.655753471  0.68948602   0   0
#>   0.76646219  0.690245059  0.76163181   0   0
#>   0.62284912  0.691013055  0.57525095   0   0
#>   0.65898565  0.816078175  0.65531269   0   0
#>   0.76163181  0.756163265  0.76182386   0   0
#>   0.68948602  0.653706321  0.68788526   0   0
#>   0.27564847  0.054801409  0.06674642   0   0
#>   0.28633689 -0.893532743  0.28687854   0   0
#>   0.56354216  0.136102918  0.54701699   0   0
#>   0.65393097  0.137770147  0.73557667   0   0
#>   0.71170741  0.092967977  0.79115186   0   0
#>   0.72124873  0.065596948  0.85537462   0   0
#>   0.78432971  0.061102947  0.88843136   0   0
#>   0.91716955  0.551264670  1.27052118   0   0
#>   0.94454206  0.358071368  0.98849823   0   0
#>   0.86680524  0.304275899  0.84953894   0   0
#>   0.84289730  0.461724928  0.79494825   0   0
#>   0.79115186  0.065596948  0.85537462   0   0
#>   0.73557667  0.207272816  0.57769039   0   0
#>   0.76444816  0.200061193  0.78443008   0   0
#>   0.59852726  0.392534978  0.58547951   0   0
#>   0.50534026  0.432166508  0.49226032   0   0
#>   0.50062967  0.527980852  0.40922075   0   0
#>   0.53599579  0.553979680  0.41440130   0   0
#>   0.54951654  0.644335392  0.46311755   0   0
#>   0.56110649  0.593965143  0.59220080   0   0
#>   0.63739418  0.461263441  0.67504563   0   0
#>   0.67384890  0.918720571  0.67231119   0   0
#>   0.69724350  0.925881804  0.71332419   0   0
#>   0.71221639  0.866894472  0.74385433   0   0
#>   0.65608612  0.816291384  0.65517948   0   0
#>   0.83162676  0.837653108  0.83610878   0   0
#>   0.84242797  0.737450926  0.91190872   0   0
#>   0.84436418  0.721433829  0.88669458   0   0
#>   0.82970889  0.684745998  0.84620006   0   0
#>   0.76182386  0.771009461  0.74725944   0   0
#>   0.94873953  0.092691795  1.27185882   0   0
#>   0.89819493  0.167169337  0.88735789   0   0
#>   0.89378586  0.304211279  0.84954553   0   0
#>   0.90182129  0.320124909  0.97395119   0   0
#>   0.93603730  0.261676717  0.98148982   0   0
#>   0.70352521  0.211703290  0.78238585   0   0
#>   0.63028944  0.207272816  0.57769039   0   0
#>   0.63009867  0.370616116  0.57419180   0   0
#>   0.65273191  0.392534978  0.58547951   0   0
#>   0.70771276  0.442217075  0.70647796   0   0
#>   0.74495948  0.319927022  0.82856948   0   0
#>   0.33440938 -0.893532743  0.28687854   0   0
#>   0.24417394  0.146164591  0.21633534   0   0
#>   0.35380368  0.100119094  0.38019929   0   0
#>   0.35097582  0.126252131  0.51658869   0   0
#>   0.19131838 -0.893532743  0.28687854   0   0
#>   0.09794806  0.054801409  0.06674642   0   0
#>   0.08911257  0.171678133  0.04244890   0   0
#>   0.11483065  0.200788010  0.09740810   0   0
#>   0.23445234  0.209672078  0.24314209   0   0
#>   0.29606658  0.519920086  0.39742554   0   0
#>   0.25217997  0.527928187  0.25407460   0   0
#>   0.19404732  0.550288167  0.12633473   0   0
#>   0.18827991  0.753684958  0.10656279   0   0
#>   0.24566603  0.689892793  0.25841230   0   0
#>   0.29629188  0.644184478  0.37781824   0   0
#>   0.12633473  0.409127848 -0.01924239   0   0
#>   0.04188627  0.793291270 -0.09661269   0   0
#>   0.10656279  0.821503092  0.20471681   0   0
#>   0.50838274  0.370616116  0.57419180   0   0
#>   0.49876001  0.326433737  0.49888023   0   0
#>   0.38172211  0.362301634  0.34624459   0   0
#>   0.44040499  0.523941040  0.40619989   0   0
#>   0.41361310  0.326433737  0.49888023   0   0
#>   0.42893147  0.195170631  0.55581858   0   0
#>   0.41205350  0.107858600  0.37556265   0   0
#>   0.24739516  0.209672078  0.24314209   0   0
#>   0.24289399  0.393768491  0.01738640   0   0
#>   0.26108115  0.489601840  0.23499807   0   0
#>   0.33801953  0.362301634  0.34624459   0   0
#>   0.40619989  0.519920086  0.39742554   0   0
#>   0.40922075  0.553979680  0.41440130   0   0
#>   0.65531269  0.816291384  0.65517948   0   0
#>   0.74725944  0.866894472  0.74385433   0   0
#>   0.84953894  0.304211279  0.84954553   0   0
#>   0.79494825  0.462604815  0.76280433   0   0
#>   0.88454209  0.888785266  0.77817948   0   0
#>   0.71342360  0.927816433  0.71954679   0   0
#>   0.70623534  0.919271338  0.67147442   0   0
#>   0.55639568  0.919271338  0.67147442   0   0
#>   0.52251257  0.918330954  0.50658899   0   0
#>   0.25361079  1.066015736  0.22678270   0   0
#>   0.98148982  0.092691795  1.27185882   0   0
#>   0.97395119  0.358071368  0.98849823   0   0
#>   0.67147442  0.918720571  0.67231119   0   0
#>   0.68201484  0.461263441  0.67504563   0   0
#>   0.66754116  0.595723436  0.62587511   0   0
#>   0.68788526  0.595723436  0.62587511   0   0
#>   0.57525095  0.593965143  0.59220080   0   0
#>   0.53622032  0.695429569  0.50445646   0   0
#>   0.55130042  0.918330954  0.50658899   0   0
#>   0.47800926  0.767449325  0.41544145   0   0
#>   0.47750062  0.820092908  0.30824757   0   0
#>   0.28571258  0.835621424  0.27357922   0   0
#>   0.22678270  0.966793375  0.16548386   0   0
#>   0.33801009  0.519920086  0.39742554   0   0
#>   0.23499807  0.397449726  0.01151570   0   0
#>   0.84620006  0.721433829  0.88669458   0   0
#>   0.93333860  0.665114017  0.94597182   0   0
#>   0.92212777  0.616407246  0.98780727   0   0
#>   0.57769039  0.195111115  0.55588660   0   0
#>   0.78238585  0.304211279  0.84954553   0   0
#>   0.78443008  0.167169337  0.88735789   0   0
#>  -0.09661269  0.966793375  0.16548386   0   0
#>  -0.01924239  0.397449726  0.01151570   0   0
#>  -1.10887009  0.257536890 -0.42851116   0   0
#>  -3.74296280  0.931037225 -4.76845703   0   1
#>   0.25841230  0.833278709  0.26739271   0   0
#>   0.34383868  0.739118709  0.35106341   0   0
#>   0.37781824  0.626545239  0.41479963   0   0
#>   0.04244890  0.257536890 -0.42851116   0   0
#>   0.09740810  0.393768491  0.01738640   0   0
#>   0.26739271  0.821503092  0.20471681   0   0
#>   0.35106341  0.767449325  0.41544145   0   0
#>   0.98366702  0.665114017  0.94597182   0   0
#>   0.94678947  0.743531769  0.92946717   0   0
#>   0.94193996  0.847183593  0.86961569   0   0
#>   0.83610878  0.888785266  0.77817948   0   0
#>   0.91190872  0.721433829  0.88669458   0   0
#>   1.27052118  0.616407246  0.98780727   0   0
#>   0.38019929  0.107858600  0.37556265   0   0
#>   0.51658869  0.136102918  0.54701699   0   0
#>   0.41479963  0.553979680  0.41440130   0   0
#>   0.71954679  0.888168929  0.77453131   0   0
#>   0.49888023  0.195170631  0.55581858   0   0
#>   0.85537462  0.061102947  0.88843136   0   0
#>   0.90503924  0.167169337  0.88735789   0   0
#>   1.27192486  0.092691795  1.27185882   0   0
#>   0.54701699  0.195111115  0.55588660   0   0
#>   0.94597182  0.616407246  0.98780727   0   0
#>   0.01738640  0.397449726  0.01151570   0   0
#>   0.82580408  0.576366828  0.90448982   0   0
#>   0.90448982  0.509734330  0.90240751   0   0
#>   0.90240751  0.604578422  0.77390653   0   0
#>   0.77390653  0.629712892  0.76171218   0   0
#>   0.76171218  0.661131864  0.76016211   0   0
#>   0.76016211  0.669710126  0.76646219   0   0
#>   0.76646219  0.675832937  0.82580408   0   0
#>   0.62284912  0.755845207  0.65898565   0   0
#>   0.65898565  0.690245059  0.76163181   0   0
#>   0.68948602  0.693382738  0.62284912   0   0
#>   0.27564847 -0.919744069  0.28633689   0   0
#>   0.28633689  0.100983779  0.56354216   0   0
#>   0.56354216  0.061378033  0.65393097   0   0
#>   0.65393097 -0.028098070  0.71170741   0   0
#>   0.71170741 -0.095684446  0.72124873   0   0
#>   0.72124873 -4.465265278  0.78432971   0   0
#>   0.78432971 -7.455328057  0.78729798   0   1
#>   0.91716955  0.381166960  0.94454206   0   0
#>   0.94454206  0.329832830  0.86680524   0   0
#>   0.86680524  0.470620411  0.84289730   0   0
#>   0.84289730  0.505417586  0.90375672   0   0
#>   0.90375672  0.499546553  0.91716955   0   0
#>   0.73557667  0.145713671  0.76444816   0   0
#>   0.76444816  0.092967977  0.79115186   0   0
#>   0.59852726  0.444494823  0.50534026   0   0
#>   0.50534026  0.494332856  0.50062967   0   0
#>   0.50062967  0.541857561  0.53599579   0   0
#>   0.53599579  0.553301440  0.54951654   0   0
#>   0.54951654  0.553219819  0.56110649   0   0
#>   0.56110649  0.461653126  0.63739418   0   0
#>   0.63739418  0.434407968  0.59852726   0   0
#>   0.67384890  0.885714528  0.69724350   0   0
#>   0.69724350  0.858592553  0.71221639   0   0
#>   0.71221639  0.817189794  0.65608612   0   0
#>   0.65608612  0.914078033  0.67384890   0   0
#>   0.83162676  0.797299779  0.84242797   0   0
#>   0.84242797  0.779462078  0.84436418   0   0
#>   0.84436418  0.678301552  0.82970889   0   0
#>   0.76182386  0.824943009  0.83162676   0   0
#>   0.94873953  0.177866316  0.89819493   0   0
#>   0.89819493  0.275213733  0.89378586   0   0
#>   0.89378586  0.273866872  0.90182129   0   0
#>   0.90182129  0.255419926  0.93603730   0   0
#>   0.93603730  0.191588477  0.94873953   0   0
#>   0.70352521  0.257654328  0.63028944   0   0
#>   0.63028944  0.264497668  0.63009867   0   0
#>   0.63009867  0.334401868  0.65273191   0   0
#>   0.65273191  0.428528364  0.70771276   0   0
#>   0.70771276  0.355226785  0.74495948   0   0
#>   0.74495948  0.258932842  0.70352521   0   0
#>   0.33440938  0.111179431  0.24417394   0   0
#>   0.24417394 -0.005419819  0.35380368   0   0
#>   0.35380368 -0.109572715  0.35097582   0   0
#>   0.35097582 -0.290069902  0.33440938   0   0
#>   0.19131838  0.091757666  0.09794806   0   0
#>   0.09794806  0.118814702  0.08911257   0   0
#>   0.08911257  0.193413959  0.11483065   0   0
#>   0.11483065  0.198757699  0.23445234   0   0
#>   0.23445234  0.146164591  0.21633534   0   0
#>   0.21633534  0.111751121  0.19131838   0   0
#>   0.29606658  0.535241871  0.25217997   0   0
#>   0.25217997  0.577433715  0.19404732   0   0
#>   0.19404732  0.680522048  0.18827991   0   0
#>   0.18827991  0.685024856  0.24566603   0   0
#>   0.24566603  0.583343425  0.29629188   0   0
#>   0.29629188  0.583011504  0.29606658   0   0
#>   0.12633473  0.759663162  0.04188627   0   0
#>   0.04188627  0.753684958  0.10656279   0   0
#>   0.50838274  0.327348107  0.49876001   0   0
#>   0.49876001  0.372239935  0.38172211   0   0
#>   0.38172211  0.428745038  0.44040499   0   0
#>   0.44040499  0.432166508  0.49226032   0   0
#>   0.49226032  0.344006275  0.50838274   0   0
#>   0.41361310  0.227322422  0.42893147   0   0
#>   0.42893147  0.217129298  0.41205350   0   0
#>   0.41205350  0.233183319  0.24739516   0   0
#>   0.24739516  0.298861620  0.24289399   0   0
#>   0.24289399  0.316667327  0.26108115   0   0
#>   0.26108115  0.353304889  0.33801953   0   0
#>   0.33801953  0.273823827  0.41361310   0   0
#>   0.40619989  0.527980852  0.40922075   0   0
#>   0.65531269  0.771009461  0.74725944   0   0
#>   0.84953894  0.319927022  0.82856948   0   0
#>   0.82856948  0.461724928  0.79494825   0   0
#>   0.88454209  0.980248136  0.71342360   0   0
#>   0.71342360  0.978016095  0.70623534   0   0
#>   0.70623534  0.962169674  0.55639568   0   0
#>   0.55639568  0.967500502  0.52251257   0   0
#>   0.52251257  6.194117272  0.25361079   0   0
#>   0.25361079  7.888630603  0.21204627   0   1
#>   0.97395119  0.261676717  0.98148982   0   0
#>   0.68201484  0.556569004  0.66754116   0   0
#>   0.66754116  0.653706321  0.68788526   0   0
#>   0.57525095  0.713223321  0.53622032   0   0
#>   0.53622032  0.845077089  0.55130042   0   0
#>   0.55130042  0.816291384  0.65517948   0   0
#>   0.50658899  0.905475878  0.47800926   0   0
#>   0.47800926  0.905420249  0.47750062   0   0
#>   0.47750062  0.956038841  0.28571258   0   0
#>   0.28571258  1.066015736  0.22678270   0   0
#>   0.33801009  0.362301634  0.34624459   0   0
#>   0.23499807  0.527928187  0.25407460   0   0
#>   0.25407460  0.483150554  0.33801009   0   0
#>   0.76280433  0.442217075  0.70647796   0   0
#>   0.70647796  0.461263441  0.67504563   0   0
#>   0.84620006  0.655191674  0.93333860   0   0
#>   0.93333860  0.601723425  0.92212777   0   0
#>   0.78238585  0.200061193  0.78443008   0   0
#>  -0.01924239  0.468998034 -1.10887009   0   0
#>  -1.10887009  0.839556948 -3.74296280   0   0
#>  -3.74296280  0.793291270 -0.09661269   0   0
#>   0.25841230  0.699276470  0.34383868   0   0
#>   0.34383868  0.644184478  0.37781824   0   0
#>   0.04244890  0.200788010  0.09740810   0   0
#>   0.26739271  0.835621424  0.27357922   0   0
#>   0.27357922  0.820092908  0.30824757   0   0
#>   0.30824757  0.739118709  0.35106341   0   0
#>   0.98366702  0.742151784  0.94678947   0   0
#>   0.94678947  0.898547853  0.94193996   0   0
#>   0.94193996  5.429439559  4.99211527   0   1
#>   0.83610878  0.847183593  0.86961569   0   0
#>   0.86961569  0.743531769  0.92946717   0   0
#>   0.92946717  0.737450926  0.91190872   0   0
#>   1.27052118  0.807943901  5.76564450   0   1
#>   0.38019929  0.126252131  0.51658869   0   0
#>   0.41544145  0.695429569  0.50445646   0   0
#>   0.50445646  0.644335392  0.46311755   0   0
#>   0.46311755  0.626545239  0.41479963   0   0
#>   0.71954679  0.925881804  0.71332419   0   0
#>   0.71332419  0.918720571  0.67231119   0   0
#>   0.59220080  0.595723436  0.62587511   0   0
#>   0.74385433  0.888168929  0.77453131   0   0
#>   0.77453131  0.888785266  0.77817948   0   0
#>   0.24314209  0.107858600  0.37556265   0   0
#>  -0.42851116  0.054801409  0.06674642   0   0
#>   0.16548386  0.821503092  0.20471681   0   0
#>   0.58547951  0.370616116  0.57419180   0   0
#>   0.88843136  0.073881543  0.90503924   0   0
#>   0.90503924  0.092663513  1.27192486   0   0
#>   1.27192486 -0.505733168  4.58212956   0   1
#>   0.55588660  0.195170631  0.55581858   0   0
#>   0.27564847 -2.660990679  0.22749223   0   1
#>   0.88454209  5.879301035  3.31963167   0   1
#>   0.98366702  1.249470574  7.30745280   0   1
#>   0.98849823  0.790538519  8.58056575   0   1
#> 
#> $x
#>                [,1]       [,2]
#>   [1,] 0.5972842675 0.84240939
#>   [2,] 0.7205058203 0.67462929
#>   [3,] 0.0000000000 0.61267643
#>   [4,] 0.3987328284 0.88256028
#>   [5,] 0.0737224347 0.72684747
#>   [6,] 0.4856343048 0.54518488
#>   [7,] 0.8856405630 0.67927441
#>   [8,] 0.7561173224 0.82602123
#>   [9,] 0.2215783580 0.92031872
#>  [10,] 0.3291978092 0.69285374
#>  [11,] 0.0579291596 0.26086098
#>  [12,] 0.1521424482 0.11951586
#>  [13,] 0.6220286070 0.25683321
#>  [14,] 0.6147469623 0.12667896
#>  [15,] 0.3828484546 0.46473455
#>  [16,] 0.2528731331 0.37750341
#>  [17,] 0.4774403959 0.45849327
#>  [18,] 0.7564913888 0.69762731
#>  [19,] 0.3896694936 0.82918869
#>  [20,] 1.0000000000 0.67608666
#>  [21,] 0.2871507781 0.95567091
#>  [22,] 0.9502122337 0.68135201
#>  [23,] 0.5894986266 0.68459883
#>  [24,] 0.7517913868 0.62055178
#>  [25,] 0.9840677985 0.36641406
#>  [26,] 0.4472611092 0.28493711
#>  [27,] 0.5462584992 0.73723729
#>  [28,] 0.6415949106 0.89842217
#>  [29,] 0.1883382096 0.69531280
#>  [30,] 0.5964702347 0.08136506
#>  [31,] 0.6381457667 0.28920434
#>  [32,] 0.1637980395 0.08570702
#>  [33,] 0.3819671007 0.79670376
#>  [34,] 0.7470900156 0.27723734
#>  [35,] 0.8029530476 0.96684108
#>  [36,] 0.7781499711 0.88240873
#>  [37,] 0.7605967434 0.86728885
#>  [38,] 0.4241587475 0.99252130
#>  [39,] 0.0527925782 0.45004688
#>  [40,] 0.7206605761 0.42298787
#>  [41,] 0.9428221056 0.69384107
#>  [42,] 0.5566768822 0.63045595
#>  [43,] 0.5412672843 0.80106434
#>  [44,] 0.7698363334 0.46277471
#>  [45,] 0.8046203971 0.38603997
#>  [46,] 0.3088384173 0.43634678
#>  [47,] 0.0263164010 0.79908991
#>  [48,] 0.8203543144 0.76272581
#>  [49,] 0.9126161216 0.84416206
#>  [50,] 0.0912461441 0.29629604
#>  [51,] 0.4547920658 0.39546084
#>  [52,] 0.1887323849 0.37124975
#>  [53,] 0.0180922888 0.00000000
#>  [54,] 0.2994946916 0.76188429
#>  [55,] 0.8865816942 0.15180529
#>  [56,] 0.3957822850 0.53545901
#>  [57,] 0.0559119431 0.25189104
#>  [58,] 0.8191328940 0.72833174
#>  [59,] 0.0003320027 0.94711724
#>  [60,] 0.0387269613 0.47007565
#>  [61,] 0.6872243966 0.91389809
#>  [62,] 0.3256012725 0.56383665
#>  [63,] 0.9008808487 0.70688075
#>  [64,] 0.8918356175 0.34207124
#>  [65,] 0.9483415576 0.74110308
#>  [66,] 0.2349388380 0.11581718
#>  [67,] 0.5935759512 0.66513076
#>  [68,] 0.6210228704 0.54613833
#>  [69,] 0.3172661533 0.93635745
#>  [70,] 0.2193994191 0.87221021
#>  [71,] 0.2784363936 0.45510486
#>  [72,] 0.3537011038 0.61717375
#>  [73,] 0.6350347309 0.62636453
#>  [74,] 0.9043695325 0.21767850
#>  [75,] 0.5345130893 0.47950197
#>  [76,] 0.1277764854 0.04489962
#>  [77,] 0.6287146228 0.95985265
#>  [78,] 0.8894146739 0.65868827
#>  [79,] 0.0030942567 0.82701368
#>  [80,] 0.2928683309 1.00000000
#>  [81,] 0.8015939189 0.92300937
#>  [82,] 0.5609531412 0.34681124
#>  [83,] 0.5039174253 0.17110982
#>  [84,] 0.5472187183 0.33826217
#>  [85,] 0.2301740789 0.96351407
#>  [86,] 0.5933671197 0.96775521
#>  [87,] 0.3667933280 0.62849070
#>  [88,] 0.9855332303 0.08630157
#>  [89,] 0.6458990470 0.97985955
#>  [90,] 0.1333137421 0.67109063
#>  [91,] 0.1513435207 0.93938651
#>  [92,] 0.0454993380 0.14235142
#>  [93,] 0.2118906512 0.06023412
#>  [94,] 0.1334025267 0.84472883
#>  [95,] 0.1699122543 0.42760604
#>  [96,] 0.5602107084 0.48206384
#>  [97,] 0.9033413516 0.70073476
#>  [98,] 0.7518715944 0.85532806
#>  [99,] 0.4344154721 0.19976869
#> [100,] 0.7489918901 0.24687103
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.5972843,0.8424094) [7]: 2 8 23 27 28 43 86 
#> 2: (0.7205058,0.6746293) [6]: 1 8 18 23 24 73 
#> 3: (0,0.6126764) [8]: 5 47 53 59 60 79 90 92 
#> 4: (0.3987328,0.8825603) [5]: 19 38 43 69 86 
#> 5: (0.07372243,0.7268475) [5]: 3 29 47 90 94 
#> 6: (0.4856343,0.5451849) [7]: 17 42 56 68 75 87 96 
#> 7: (0.8856406,0.6792744) [4]: 58 63 78 97 
#> 8: (0.7561173,0.8260212) [8]: 1 2 18 28 36 37 48 98 
#> 9: (0.2215784,0.9203187) [5]: 21 69 70 85 91 
#> 10: (0.3291978,0.6928537) [6]: 29 33 54 62 72 87 
#> 11: (0.05792916,0.260861) [4]: 39 50 57 60 
#> 12: (0.1521424,0.1195159) [6]: 32 50 57 66 76 92 
#> 13: (0.6220286,0.2568332) [6]: 14 31 82 83 84 100 
#> 14: (0.614747,0.126679) [5]: 13 30 55 83 100 
#> 15: (0.3828485,0.4647346) [5]: 17 46 51 56 62 
#> 16: (0.2528731,0.3775034) [7]: 26 46 52 66 71 95 99 
#> 17: (0.4774404,0.4584933) [6]: 6 15 51 56 75 82 
#> 18: (0.7564914,0.6976273) [5]: 2 8 24 48 58 
#> 19: (0.3896695,0.8291887) [5]: 4 33 43 54 69 
#> 20: (1,0.6760867) [7]: 22 25 41 49 65 78 88 
#> 21: (0.2871508,0.9556709) [4]: 9 69 80 85 
#> 22: (0.9502122,0.681352) [3]: 20 41 78 
#> 23: (0.5894986,0.6845988) [6]: 1 2 27 42 67 73 
#> 24: (0.7517914,0.6205518) [7]: 2 18 44 58 68 73 78 
#> 25: (0.9840678,0.3664141) [7]: 20 44 45 64 74 78 88 
#> 26: (0.4472611,0.2849371) [6]: 16 46 51 83 84 99 
#> 27: (0.5462585,0.7372373) [6]: 1 23 33 42 43 87 
#> 28: (0.6415949,0.8984222) [6]: 1 8 61 77 86 98 
#> 29: (0.1883382,0.6953128) [7]: 5 10 54 62 70 90 94 
#> 30: (0.5964702,0.08136506) [6]: 14 53 55 83 88 93 
#> 31: (0.6381458,0.2892043) [5]: 13 34 40 82 100 
#> 32: (0.163798,0.08570702) [4]: 12 66 76 93 
#> 33: (0.3819671,0.7967038) [6]: 10 19 27 43 54 87 
#> 34: (0.74709,0.2772373) [6]: 31 40 45 64 74 100 
#> 35: (0.802953,0.9668411) [4]: 49 61 81 89 
#> 36: (0.77815,0.8824087) [6]: 8 37 48 49 61 81 
#> 37: (0.7605967,0.8672889) [4]: 8 36 61 98 
#> 38: (0.4241587,0.9925213) [5]: 4 69 80 86 89 
#> 39: (0.05279258,0.4500469) [4]: 11 50 60 95 
#> 40: (0.7206606,0.4229879) [7]: 31 34 44 45 68 82 96 
#> 41: (0.9428221,0.6938411) [6]: 20 22 63 65 78 97 
#> 42: (0.5566769,0.630456) [7]: 6 23 27 67 68 73 87 
#> 43: (0.5412673,0.8010643) [6]: 1 4 19 27 33 86 
#> 44: (0.7698363,0.4627747) [6]: 24 25 40 45 68 78 
#> 45: (0.8046204,0.38604) [5]: 25 34 40 44 64 
#> 46: (0.3088384,0.4363468) [6]: 15 16 26 51 62 71 
#> 47: (0.0263164,0.7990899) [4]: 3 5 79 94 
#> 48: (0.8203543,0.7627258) [7]: 8 18 36 49 58 63 65 
#> 49: (0.9126161,0.8441621) [6]: 20 35 36 48 65 81 
#> 50: (0.09124614,0.296296) [7]: 11 12 39 52 57 66 95 
#> 51: (0.4547921,0.3954608) [6]: 15 17 26 46 82 84 
#> 52: (0.1887324,0.3712498) [4]: 16 50 66 95 
#> 53: (0.01809229,0) [6]: 3 30 76 88 92 93 
#> 54: (0.2994947,0.7618843) [6]: 10 19 29 33 69 70 
#> 55: (0.8865817,0.1518053) [5]: 14 30 74 88 100 
#> 56: (0.3957823,0.535459) [6]: 6 15 17 62 72 87 
#> 57: (0.05591194,0.251891) [5]: 11 12 50 60 92 
#> 58: (0.8191329,0.7283317) [6]: 7 18 24 48 63 78 
#> 59: (0.0003320027,0.9471172) [5]: 3 79 80 91 94 
#> 60: (0.03872696,0.4700757) [7]: 3 11 39 57 90 92 95 
#> 61: (0.6872244,0.9138981) [8]: 28 35 36 37 77 81 89 98 
#> 62: (0.3256013,0.5638366) [9]: 10 15 29 46 56 71 72 90 95 
#> 63: (0.9008808,0.7068807) [6]: 7 41 48 58 65 97 
#> 64: (0.8918356,0.3420712) [4]: 25 34 45 74 
#> 65: (0.9483416,0.7411031) [5]: 20 41 48 49 63 
#> 66: (0.2349388,0.1158172) [7]: 12 16 32 50 52 93 99 
#> 67: (0.593576,0.6651308) [3]: 23 42 73 
#> 68: (0.6210229,0.5461383) [7]: 6 24 40 42 44 73 96 
#> 69: (0.3172662,0.9363574) [8]: 4 9 19 21 38 54 70 80 
#> 70: (0.2193994,0.8722102) [6]: 9 29 54 69 91 94 
#> 71: (0.2784364,0.4551049) [4]: 16 46 62 95 
#> 72: (0.3537011,0.6171738) [4]: 10 56 62 87 
#> 73: (0.6350347,0.6263645) [6]: 2 23 24 42 67 68 
#> 74: (0.9043695,0.2176785) [6]: 25 34 55 64 88 100 
#> 75: (0.5345131,0.479502) [4]: 6 17 82 96 
#> 76: (0.1277765,0.04489962) [5]: 12 32 53 92 93 
#> 77: (0.6287146,0.9598526) [4]: 28 61 86 89 
#> 78: (0.8894147,0.6586883) [9]: 7 20 22 24 25 41 44 58 97 
#> 79: (0.003094257,0.8270137) [4]: 3 47 59 94 
#> 80: (0.2928683,1) [6]: 21 38 59 69 85 91 
#> 81: (0.8015939,0.9230094) [4]: 35 36 49 61 
#> 82: (0.5609531,0.3468112) [8]: 13 17 31 40 51 75 84 96 
#> 83: (0.5039174,0.1711098) [7]: 13 14 26 30 84 93 99 
#> 84: (0.5472187,0.3382622) [5]: 13 26 51 82 83 
#> 85: (0.2301741,0.9635141) [4]: 9 21 80 91 
#> 86: (0.5933671,0.9677552) [7]: 1 4 28 38 43 77 89 
#> 87: (0.3667933,0.6284907) [7]: 6 10 27 33 42 56 72 
#> 88: (0.9855332,0.08630157) [6]: 20 25 30 53 55 74 
#> 89: (0.645899,0.9798595) [5]: 35 38 61 77 86 
#> 90: (0.1333137,0.6710906) [6]: 3 5 29 60 62 95 
#> 91: (0.1513435,0.9393865) [6]: 9 59 70 80 85 94 
#> 92: (0.04549934,0.1423514) [6]: 3 12 53 57 60 76 
#> 93: (0.2118907,0.06023412) [7]: 30 32 53 66 76 83 99 
#> 94: (0.1334025,0.8447288) [7]: 5 29 47 59 70 79 91 
#> 95: (0.1699123,0.427606) [8]: 16 39 50 52 60 62 71 90 
#> 96: (0.5602107,0.4820638) [5]: 6 40 68 75 82 
#> 97: (0.9033414,0.7007348) [4]: 7 41 63 78 
#> 98: (0.7518716,0.8553281) [4]: 8 28 37 61 
#> 99: (0.4344155,0.1997687) [5]: 16 26 66 83 93 
#> 100: (0.7489919,0.246871) [6]: 13 14 31 34 55 74 
#> number of nodes: 100 
#> number of arcs: 287 
#> number of boundary nodes: 10 
#> boundary nodes:  3 20 35 38 49 53 59 80 88 89 
#> number of triangles: 188 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.135489039 0.084457029 0.174643178 0.064021901 0.068121335 0.131385983 
#>                                                                         
#> 0.155523946 0.116901261 0.077218705 0.128394461 0.073924091 0.144965768 
#>                                                                         
#> 0.327794196 0.222156885 0.081608510 0.086407770 0.036318103 0.120135320 
#>                                                                         
#> 0.171011206 0.120749349 0.129334400 0.154184529 0.116405968 0.060119924 
#>                                                                         
#> 0.159195107 0.097443595 0.112213063 0.060816589 0.025825004 0.088338325 
#>                                                                         
#> 0.106065393 0.189893724 0.044292882 0.006620215 0.084515519 0.098943117 
#>                                                                         
#> 0.126906299 0.023167360 0.014805026 0.118397845 0.091193681 0.082440258 
#>                                                                         
#> 0.095625628 0.117016093 0.035776330 0.057513993 0.129566689 0.190071372 
#>                                                                         
#> 0.060286378 0.017305478 0.168896056 0.089521472 0.218860337 0.056747769 
#>                                                                         
#> 0.158486589 0.024474338 0.110033406 0.127539729 0.054431572 0.077250468 
#>                                                                         
#> 0.230695057 0.016177789 0.172669948 0.119403884 0.180188325 0.118654845 
#>                                                                         
#> 0.096318516 0.128919143 0.298540593 0.167237698 0.075701154 0.128587170 
#>                                                                         
#> 0.151572184 0.066977865 0.035723223 0.111953891 0.059415724 0.259578193 
#>                                                                         
#> 0.216422698 0.086131702 0.205148065 0.116777386 0.135299318 0.127088152 
#>                                                                         
#> 0.034415753 0.175375907 0.159359854 0.109075480 0.047583213 0.014511741 
#>                                                                         
#> 0.064884422 0.307218861 0.280116323 0.072538213 0.068158846 0.063938039 
#>                                                                         
#> 0.107288384 0.050634803 0.056759591 0.081440619 0.170572388 0.229523642 
#>                                                                         
#> 0.084250529 0.097671616 0.125022600 0.154426177 0.108693945 0.075178802 
#>                                                                         
#> 0.087233643 0.074399026 0.036220112 0.220177001 0.136334373 0.090281179 
#>                                                                         
#> 0.118668210 0.312366864 0.202943231 0.971282625 0.030425809 0.148127427 
#>                                                                         
#> 0.176944467 0.085500525 0.060172209 0.158096221 0.091730033 0.077837645 
#>                                                                         
#> 0.114731874 0.136172047 0.123061386 0.086924808 0.053908426 0.153070813 
#>                                                                         
#> 0.137888538 0.135254638 0.058512281 0.118520580 0.131506928 0.096342899 
#>                                                                         
#> 0.153958655 0.246219816 0.026373862 0.262654129 0.071420339 0.125407009 
#>                                                                         
#> 0.069622668 0.116896533 0.158002499 0.208167513 0.159676270 0.062475310 
#>                                                                         
#> 0.042706792 0.098156990 0.472520689 0.147765895 0.145549898 0.135904466 
#>                                                                         
#> 0.188261892 0.214359583 0.334440974 0.112862319 0.097626606 0.054135674 
#>                                                                         
#> 0.164187880 0.212463369 0.118874772 0.132127702 0.090376866 0.087077975 
#>                                                                         
#> 0.081874173 0.097703149 0.135391923 0.110987407 0.145131105 0.027818422 
#>                                                                         
#> 0.031533733 0.082643157 0.020929235 0.060539145 0.041510022 0.029612777 
#>                                                                         
#> 0.090181492 0.072777146 0.048157832 0.097022652 0.074495099 0.044042306 
#>                                                                         
#> 0.140881062 0.129067211 0.079547920 0.074538742 0.116487875 0.075149796 
#>                                                                         
#> 0.009193967 0.048638096 0.189255614 0.210094029 0.109060607 0.078493837 
#>                                                                         
#> 0.035761575 0.082878962 0.186974843 0.163656653 0.110576648 0.145940917 
#>                                                                         
#> 0.130357783 0.127353526 0.036161477 0.108748597 0.048860913 0.272993505 
#>                                                                         
#> 0.114448525 0.079267594 0.099873571 0.094797621 0.071897377 0.081703516 
#>                                                                         
#> 0.096916359 0.064444890 0.262300067 0.254061512 0.215302591 0.081207478 
#>                                                                         
#> 0.139453418 0.069761879 0.112522785 0.033385586 0.083040531 0.059870950 
#>                                                                         
#> 0.050065413 0.111945604 0.310082172 0.589962489 0.044696292 0.063314439 
#>                                                                         
#> 0.019890462 0.150458518 0.158805609 0.142809529 0.234905291 0.180517440 
#>                                                                         
#> 0.095390505 0.168742643 0.110780010 0.127147904 0.113292039 0.174722420 
#>                                                                         
#> 0.209841774 0.048182506 0.062766271 0.179603710 0.385159671 0.584073044 
#>                                                                         
#> 0.389094312 0.109599537 0.157183741 0.054422126 0.168178776 0.158602408 
#>                                                                         
#> 0.123076231 0.127263823 0.043852778 0.164548261 0.139799690 0.046883154 
#>                                                                         
#> 0.096223904 0.222101509 0.119250199 0.063255417 0.158409936 0.170979910 
#>                                                                         
#> 0.043921547 0.040078085 0.078464592 0.097995885 0.129800912 0.122970025 
#>                                                                         
#> 0.118518349 0.068232596 0.091913682 0.167902202 0.151209268 0.297277799 
#>                                                             
#> 0.206876409 0.612943503 0.189434112 0.157592639 0.131503249 
#> 
#> $alpha
#> [1] 0.2855074
#> 
#> attr(,"class")
#> [1] "scree"