Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2          mx1
#>    39   52 0.583097386 0.957640422 0.553834020 0.864202436  0.547817934
#>    29   65 0.444049320 0.994817170 0.407785633 0.921020138  0.456294029
#>    65   52 0.407785633 0.921020138 0.553834020 0.864202436  0.486251856
#>    39    1 0.583097386 0.957640422 0.601328297 0.916147676  0.666616803
#>    49   93 0.662553387 0.787711691 0.718040065 0.902698714  0.654286543
#>    52   49 0.553834020 0.864202436 0.662553387 0.787711691  0.623704775
#>    65   81 0.407785633 0.921020138 0.344585822 0.844090935  0.433234009
#>    59   81 0.314642350 0.871850793 0.344585822 0.844090935  0.362686923
#>    45   68 0.400603455 0.731749485 0.584070790 0.645133186  0.513700991
#>    68   52 0.584070790 0.645133186 0.553834020 0.864202436  0.522254855
#>     9   81 0.402120197 0.752255743 0.344585822 0.844090935  0.365804266
#>    49   55 0.662553387 0.787711691 0.670419076 0.714406241  0.717045025
#>    68   55 0.584070790 0.645133186 0.670419076 0.714406241  0.577696212
#>    45    9 0.400603455 0.731749485 0.402120197 0.752255743  0.365848194
#>    98   93 0.778257492 0.839498251 0.718040065 0.902698714  0.710659861
#>    43    4 0.296864447 0.887090815 0.278682859 0.889736188  0.282531066
#>    43   59 0.296864447 0.887090815 0.314642350 0.871850793  0.345619819
#>    98   49 0.778257492 0.839498251 0.662553387 0.787711691  0.730257584
#>    20    4 0.243700563 0.956784860 0.278682859 0.889736188  0.295443977
#>     4   27 0.278682859 0.889736188 0.195207658 0.798362838  0.265439638
#>    45  100 0.400603455 0.731749485 0.309897352 0.734239018  0.353564070
#>    16   49 0.793319502 0.816535072 0.662553387 0.787711691  0.735869385
#>    15   26 0.663433266 0.630241367 0.710583225 0.694011888  0.674011972
#>    29   73 0.444049320 0.994817170 0.487176380 0.986374795  3.394873172
#>    73   39 0.487176380 0.986374795 0.583097386 0.957640422  0.864590038
#>    39   93 0.583097386 0.957640422 0.718040065 0.902698714  0.736624341
#>    98   16 0.778257492 0.839498251 0.793319502 0.816535072  0.836265562
#>    36  100 0.329562524 0.752190449 0.309897352 0.734239018  0.354485932
#>    59   27 0.314642350 0.871850793 0.195207658 0.798362838  0.272904380
#>    20   78 0.243700563 0.956784860 0.200938254 0.963335115  0.212816377
#>    43   65 0.296864447 0.887090815 0.407785633 0.921020138  0.332368229
#>    68   15 0.584070790 0.645133186 0.663433266 0.630241367  0.630813770
#>    68   50 0.584070790 0.645133186 0.531107659 0.600365866  0.597870200
#>    20   65 0.243700563 0.956784860 0.407785633 0.921020138  0.338751242
#>    31   76 0.751255589 0.648520506 0.845111679 0.652894085  0.794631573
#>    16   26 0.793319502 0.816535072 0.710583225 0.694011888  0.793499649
#>    88   31 0.669130206 0.579326284 0.751255589 0.648520506  0.713446973
#>    50   45 0.531107659 0.600365866 0.400603455 0.731749485  0.469084567
#>    27  100 0.195207658 0.798362838 0.309897352 0.734239018  0.270064227
#>    20   75 0.243700563 0.956784860 0.234926034 1.000000000  0.224672175
#>    72   82 0.475643188 0.489848100 0.572821313 0.406000955  0.576443561
#>    44   66 0.398183391 0.609540808 0.483261480 0.558324254  0.466104905
#>    78   11 0.200938254 0.963335115 0.129306338 0.884683824  0.203945867
#>    84   14 0.900223944 0.761698883 0.834796540 0.923016771  0.875807582
#>    76   84 0.845111679 0.652894085 0.900223944 0.761698883  0.818798530
#>    28   53 0.350198677 0.573927215 0.407261598 0.544323536  0.386819703
#>    28   44 0.350198677 0.573927215 0.398183391 0.609540808  0.328251570
#>    88   12 0.669130206 0.579326284 0.818452135 0.602493701  0.746998153
#>    53   72 0.407261598 0.544323536 0.475643188 0.489848100  0.449668855
#>    53   66 0.407261598 0.544323536 0.483261480 0.558324254  0.439598355
#>    12   76 0.818452135 0.602493701 0.845111679 0.652894085  0.798434050
#>    28  100 0.350198677 0.573927215 0.309897352 0.734239018  0.305776705
#>    32   27 0.220269124 0.660056151 0.195207658 0.798362838  0.155524292
#>    11   27 0.129306338 0.884683824 0.195207658 0.798362838  0.204450343
#>    22   88 0.674518335 0.458796311 0.669130206 0.579326284  0.588762779
#>    50   66 0.531107659 0.600365866 0.483261480 0.558324254  0.564005287
#>    37   10 0.111088633 0.923168323 0.000000000 0.960338310  0.028337636
#>    61   37 0.090943273 0.973153295 0.111088633 0.923168323  0.145757672
#>    18   76 0.868500737 0.594745014 0.845111679 0.652894085  0.846751890
#>    64   18 0.827548309 0.571137371 0.868500737 0.594745014  0.842525881
#>    64   12 0.827548309 0.571137371 0.818452135 0.602493701  0.747810549
#>    32   33 0.220269124 0.660056151 0.302047796 0.504038676  0.262554934
#>    76   24 0.845111679 0.652894085 0.911649256 0.611946998  0.908677455
#>    75   61 0.234926034 1.000000000 0.090943273 0.973153295  0.152548930
#>    62   69 0.971229201 0.712553689 1.000000000 0.791634646  0.954037697
#>    18   24 0.868500737 0.594745014 0.911649256 0.611946998  0.878443221
#>    63   64 0.855930571 0.540267194 0.827548309 0.571137371  0.795172851
#>    64   88 0.827548309 0.571137371 0.669130206 0.579326284  0.745606200
#>    53   23 0.407261598 0.544323536 0.346143712 0.515839324  0.395652012
#>    33   95 0.302047796 0.504038676 0.434637373 0.449060452  0.346881865
#>    33   23 0.302047796 0.504038676 0.346143712 0.515839324  0.314107857
#>    72   97 0.475643188 0.489848100 0.487029549 0.422344244  0.539790141
#>    38   64 0.878222938 0.568023455 0.827548309 0.571137371  0.852649190
#>    95   97 0.434637373 0.449060452 0.487029549 0.422344244  0.470282012
#>    46   90 0.403639354 0.324366784 0.491011202 0.353055140  0.432668074
#>    22   82 0.674518335 0.458796311 0.572821313 0.406000955  0.625369779
#>    34   48 0.094636672 0.656126413 0.204418673 0.485600281  0.159971054
#>    10   34 0.000000000 0.960338310 0.094636672 0.656126413 -0.013229307
#>    25   47 0.011646144 0.997223031 0.017245808 0.989621770  0.065126838
#>    94   38 0.869792074 0.481365748 0.878222938 0.568023455  0.907844772
#>    56   94 0.815160589 0.449315232 0.869792074 0.481365748  0.821506597
#>    64   22 0.827548309 0.571137371 0.674518335 0.458796311  0.749121229
#>    94   85 0.869792074 0.481365748 0.985809812 0.374900239  1.003416070
#>    94   24 0.869792074 0.481365748 0.911649256 0.611946998  0.996374252
#>    25   10 0.011646144 0.997223031 0.000000000 0.960338310 -0.002053485
#>    75   25 0.234926034 1.000000000 0.011646144 0.997223031  0.120612803
#>    70   71 0.670201896 0.364082324 0.747552257 0.324499577  0.741623797
#>    91   82 0.605395816 0.375299735 0.572821313 0.406000955  0.579762838
#>    47   10 0.017245808 0.989621770 0.000000000 0.960338310  0.047541022
#>    48    3 0.204418673 0.485600281 0.254231892 0.357893405  0.264446565
#>     3   95 0.254231892 0.357893405 0.434637373 0.449060452  0.340923786
#>    48   33 0.204418673 0.485600281 0.302047796 0.504038676  0.238957040
#>     7   70 0.595836265 0.360040660 0.670201896 0.364082324  0.633845673
#>    90    7 0.491011202 0.353055140 0.595836265 0.360040660  0.543037490
#>    71   56 0.747552257 0.324499577 0.815160589 0.449315232  0.741758723
#>    22   91 0.674518335 0.458796311 0.605395816 0.375299735  0.645239090
#>    46   86 0.403639354 0.324366784 0.380379296 0.306089180  0.430281406
#>    97   90 0.487029549 0.422344244 0.491011202 0.353055140  0.525279279
#>     7    8 0.595836265 0.360040660 0.626720269 0.266345355  0.635245649
#>    99   79 0.081588609 0.386792806 0.110285205 0.362230271  0.145780654
#>     8   71 0.626720269 0.266345355 0.747552257 0.324499577  0.685552947
#>    46    3 0.403639354 0.324366784 0.254231892 0.357893405  0.338425164
#>    34   67 0.094636672 0.656126413 0.057108232 0.504039557  0.139471674
#>    13   67 0.006843302 0.522266934 0.057108232 0.504039557  0.057858165
#>    60   96 0.488819005 0.241442017 0.549811220 0.232597348  0.527940785
#>    10   13 0.000000000 0.960338310 0.006843302 0.522266934 -0.176883035
#>    96    8 0.549811220 0.232597348 0.626720269 0.266345355  0.566745597
#>    60   86 0.488819005 0.241442017 0.380379296 0.306089180  0.393292635
#>    96    7 0.549811220 0.232597348 0.595836265 0.360040660  0.546811350
#>    85   62 0.985809812 0.374900239 0.971229201 0.712553689  1.130496060
#>    71   85 0.747552257 0.324499577 0.985809812 0.374900239  0.868833627
#>     8   74 0.626720269 0.266345355 0.642554159 0.154471025  0.699881206
#>     3    2 0.254231892 0.357893405 0.249691779 0.283609264  0.311202852
#>    67   99 0.057108232 0.504039557 0.081588609 0.386792806  0.126061661
#>    79    2 0.110285205 0.362230271 0.249691779 0.283609264  0.181204483
#>    96   74 0.549811220 0.232597348 0.642554159 0.154471025  0.607116029
#>    80   89 0.421887029 0.055907624 0.501859345 0.017278378  0.506614290
#>    86    2 0.380379296 0.306089180 0.249691779 0.283609264  0.324762301
#>     2   40 0.249691779 0.283609264 0.117655330 0.272817227  0.185938941
#>    99   13 0.081588609 0.386792806 0.006843302 0.522266934 -0.293620663
#>    54    8 0.778347585 0.165493016 0.626720269 0.266345355  0.706675598
#>    21   80 0.304782159 0.152926316 0.421887029 0.055907624  0.412737824
#>    83   77 0.650817257 0.118435767 0.656161307 0.000000000  0.673960608
#>    17   74 0.674517468 0.136634082 0.642554159 0.154471025  0.654451859
#>    21    2 0.304782159 0.152926316 0.249691779 0.283609264  0.259336576
#>    41    6 0.790963119 0.126624382 0.900838945 0.165261818  0.839590456
#>    41   54 0.790963119 0.126624382 0.778347585 0.165493016  0.732526253
#>    71   54 0.747552257 0.324499577 0.778347585 0.165493016  0.839771625
#>    35   57 0.071465886 0.121255598 0.090447552 0.014549975  0.163428169
#>    87   40 0.193005886 0.179006573 0.117655330 0.272817227  0.168589143
#>    35   30 0.071465886 0.121255598 0.141950325 0.180000735  0.058995804
#>    89   83 0.501859345 0.017278378 0.650817257 0.118435767  0.545509797
#>    89   74 0.501859345 0.017278378 0.642554159 0.154471025  0.536412981
#>    21   87 0.304782159 0.152926316 0.193005886 0.179006573  0.229722836
#>     5   80 0.347899572 0.061781561 0.421887029 0.055907624  0.391179663
#>    85    6 0.985809812 0.374900239 0.900838945 0.165261818  1.011073307
#>    17   83 0.674517468 0.136634082 0.650817257 0.118435767  0.697617093
#>    92   89 0.331289682 0.021660458 0.501859345 0.017278378  0.414494525
#>    41   19 0.790963119 0.126624382 0.730271147 0.032520057  0.829031365
#>     5   92 0.347899572 0.061781561 0.331289682 0.021660458  0.382135631
#>    58   92 0.300012161 0.009831486 0.331289682 0.021660458  0.299558927
#>    58   77 0.300012161 0.009831486 0.656161307 0.000000000  0.445203089
#>    57   87 0.090447552 0.014549975 0.193005886 0.179006573  0.165577637
#>    40   99 0.117655330 0.272817227 0.081588609 0.386792806 -0.116590554
#>    21   58 0.304782159 0.152926316 0.300012161 0.009831486  0.273471189
#>    85   42 0.985809812 0.374900239 0.942556562 0.126683447 54.293478428
#>    42    6 0.942556562 0.126683447 0.900838945 0.165261818  0.873522078
#>    19   77 0.730271147 0.032520057 0.656161307 0.000000000  0.842644636
#>    58   87 0.300012161 0.009831486 0.193005886 0.179006573  0.196367184
#>     1   52 0.601328297 0.916147676 0.553834020 0.864202436  0.547817934
#>    73   65 0.487176380 0.986374795 0.407785633 0.921020138  0.456294029
#>    73   52 0.487176380 0.986374795 0.553834020 0.864202436  0.486251856
#>    93    1 0.718040065 0.902698714 0.601328297 0.916147676  0.666616803
#>     1   49 0.601328297 0.916147676 0.662553387 0.787711691  0.623704775
#>     9   65 0.402120197 0.752255743 0.407785633 0.921020138  0.458337133
#>     9   68 0.402120197 0.752255743 0.584070790 0.645133186  0.513700991
#>     9   52 0.402120197 0.752255743 0.553834020 0.864202436  0.522254855
#>    36   81 0.329562524 0.752190449 0.344585822 0.844090935  0.365804266
#>    26   55 0.710583225 0.694011888 0.670419076 0.714406241  0.717045025
#>    36    9 0.329562524 0.752190449 0.402120197 0.752255743  0.365848194
#>    75   29 0.234926034 1.000000000 0.444049320 0.994817170  0.339560680
#>    59    4 0.314642350 0.871850793 0.278682859 0.889736188  0.282531066
#>    65   59 0.407785633 0.921020138 0.314642350 0.871850793  0.345619819
#>    36   59 0.329562524 0.752190449 0.314642350 0.871850793  0.282428657
#>    49   68 0.662553387 0.787711691 0.584070790 0.645133186  0.556863399
#>    44  100 0.398183391 0.609540808 0.309897352 0.734239018  0.353564070
#>    26   49 0.710583225 0.694011888 0.662553387 0.787711691  0.735869385
#>    14   73 0.834796540 0.923016771 0.487176380 0.986374795  3.394873172
#>    14   39 0.834796540 0.923016771 0.583097386 0.957640422  0.864590038
#>    14   93 0.834796540 0.923016771 0.718040065 0.902698714  0.736624341
#>    14   98 0.834796540 0.923016771 0.778257492 0.839498251  0.778631098
#>    14   16 0.834796540 0.923016771 0.793319502 0.816535072  0.836265562
#>    36   27 0.329562524 0.752190449 0.195207658 0.798362838  0.272904380
#>     4   78 0.278682859 0.889736188 0.200938254 0.963335115  0.212816377
#>    55   15 0.670419076 0.714406241 0.663433266 0.630241367  0.630813770
#>    88   50 0.669130206 0.579326284 0.531107659 0.600365866  0.597870200
#>    75   65 0.234926034 1.000000000 0.407785633 0.921020138  0.338751242
#>    16   76 0.793319502 0.816535072 0.845111679 0.652894085  0.794631573
#>    31   26 0.751255589 0.648520506 0.710583225 0.694011888  0.793499649
#>    31   15 0.751255589 0.648520506 0.663433266 0.630241367  0.705447878
#>    44   45 0.398183391 0.609540808 0.400603455 0.731749485  0.469084567
#>    78   75 0.200938254 0.963335115 0.234926034 1.000000000  0.224672175
#>    88   68 0.669130206 0.579326284 0.584070790 0.645133186  0.616533429
#>    50   82 0.531107659 0.600365866 0.572821313 0.406000955  0.576443561
#>     4   11 0.278682859 0.889736188 0.129306338 0.884683824  0.203945867
#>    16   84 0.793319502 0.816535072 0.900223944 0.761698883  0.818798530
#>    44   53 0.398183391 0.609540808 0.407261598 0.544323536  0.386819703
#>    31   12 0.751255589 0.648520506 0.818452135 0.602493701  0.746998153
#>    66   72 0.483261480 0.558324254 0.475643188 0.489848100  0.449668855
#>    32  100 0.220269124 0.660056151 0.309897352 0.734239018  0.305776705
#>    34   27 0.094636672 0.656126413 0.195207658 0.798362838  0.155524292
#>    95   53 0.434637373 0.449060452 0.407261598 0.544323536  0.426469600
#>    23   28 0.346143712 0.515839324 0.350198677 0.573927215  0.370531680
#>    11   10 0.129306338 0.884683824 0.000000000 0.960338310  0.028337636
#>    37   78 0.111088633 0.923168323 0.200938254 0.963335115  0.164267602
#>    12   18 0.818452135 0.602493701 0.868500737 0.594745014  0.842525881
#>    28   33 0.350198677 0.573927215 0.302047796 0.504038676  0.262554934
#>    62   24 0.971229201 0.712553689 0.911649256 0.611946998  0.908677455
#>    78   61 0.200938254 0.963335115 0.090943273 0.973153295  0.152548930
#>    69   84 1.000000000 0.791634646 0.900223944 0.761698883  0.923579822
#>    62   76 0.971229201 0.712553689 0.845111679 0.652894085  0.904070936
#>    38   18 0.878222938 0.568023455 0.868500737 0.594745014  0.895604235
#>    56   64 0.815160589 0.449315232 0.827548309 0.571137371  0.795172851
#>    95   23 0.434637373 0.449060452 0.346143712 0.515839324  0.395652012
#>    82   97 0.572821313 0.406000955 0.487029549 0.422344244  0.539790141
#>    95   90 0.434637373 0.449060452 0.491011202 0.353055140  0.432668074
#>    50   22 0.531107659 0.600365866 0.674518335 0.458796311  0.583417940
#>    32   48 0.220269124 0.660056151 0.204418673 0.485600281  0.159971054
#>    34   11 0.094636672 0.656126413 0.129306338 0.884683824  0.076208430
#>    61   47 0.090943273 0.973153295 0.017245808 0.989621770  0.065126838
#>    63   38 0.855930571 0.540267194 0.878222938 0.568023455  0.907844772
#>    63   94 0.855930571 0.540267194 0.869792074 0.481365748  0.821506597
#>    56   22 0.815160589 0.449315232 0.674518335 0.458796311  0.749121229
#>    24   85 0.911649256 0.611946998 0.985809812 0.374900239  1.003416070
#>    38   24 0.878222938 0.568023455 0.911649256 0.611946998  0.996374252
#>    61   25 0.090943273 0.973153295 0.011646144 0.997223031  0.120612803
#>    22   71 0.674518335 0.458796311 0.747552257 0.324499577  0.741623797
#>     7   82 0.595836265 0.360040660 0.572821313 0.406000955  0.579762838
#>    61   10 0.090943273 0.973153295 0.000000000 0.960338310  0.047541022
#>    33    3 0.302047796 0.504038676 0.254231892 0.357893405  0.264446565
#>    91   70 0.605395816 0.375299735 0.670201896 0.364082324  0.633845673
#>    82   90 0.572821313 0.406000955 0.491011202 0.353055140  0.525279279
#>    70    8 0.670201896 0.364082324 0.626720269 0.266345355  0.635245649
#>    60   46 0.488819005 0.241442017 0.403639354 0.324366784  0.460751281
#>    48   79 0.204418673 0.485600281 0.110285205 0.362230271  0.145780654
#>    86    3 0.380379296 0.306089180 0.254231892 0.357893405  0.338425164
#>    48   67 0.204418673 0.485600281 0.057108232 0.504039557  0.139471674
#>    90   96 0.491011202 0.353055140 0.549811220 0.232597348  0.527940785
#>    34   13 0.094636672 0.656126413 0.006843302 0.522266934 -0.176883035
#>    21   86 0.304782159 0.152926316 0.380379296 0.306089180  0.393292635
#>    56   85 0.815160589 0.449315232 0.985809812 0.374900239  0.868833627
#>    48   99 0.204418673 0.485600281 0.081588609 0.386792806  0.126061661
#>    96   89 0.549811220 0.232597348 0.501859345 0.017278378  0.506614290
#>    40   79 0.117655330 0.272817227 0.110285205 0.362230271  0.180015535
#>    35   13 0.071465886 0.121255598 0.006843302 0.522266934 -0.293620663
#>     6   71 0.900838945 0.165261818 0.747552257 0.324499577  0.877915252
#>    17    8 0.674517468 0.136634082 0.626720269 0.266345355  0.706675598
#>    80   60 0.421887029 0.055907624 0.488819005 0.241442017  0.503961080
#>    83   74 0.650817257 0.118435767 0.642554159 0.154471025  0.654451859
#>    87    2 0.193005886 0.179006573 0.249691779 0.283609264  0.259336576
#>    54    6 0.778347585 0.165493016 0.900838945 0.165261818  0.839590456
#>    17   54 0.674517468 0.136634082 0.778347585 0.165493016  0.732526253
#>    30   57 0.141950325 0.180000735 0.090447552 0.014549975  0.163428169
#>    30   40 0.141950325 0.180000735 0.117655330 0.272817227  0.168589143
#>    51   41 0.927730467 0.039992707 0.790963119 0.126624382  0.864863587
#>    77   89 0.656161307 0.000000000 0.501859345 0.017278378  0.584324547
#>    19   83 0.730271147 0.032520057 0.650817257 0.118435767  0.697617093
#>    51   19 0.927730467 0.039992707 0.730271147 0.032520057  0.829031365
#>    80   92 0.421887029 0.055907624 0.331289682 0.021660458  0.382135631
#>    17   41 0.674517468 0.136634082 0.790963119 0.126624382  0.729964808
#>    30   87 0.141950325 0.180000735 0.193005886 0.179006573  0.165577637
#>    35   99 0.071465886 0.121255598 0.081588609 0.386792806 -0.116590554
#>    89   58 0.501859345 0.017278378 0.300012161 0.009831486  0.410065257
#>     5   58 0.347899572 0.061781561 0.300012161 0.009831486  0.273471189
#>    51   42 0.927730467 0.039992707 0.942556562 0.126683447 54.293478428
#>    51    6 0.927730467 0.039992707 0.900838945 0.165261818  0.873522078
#>    51   77 0.927730467 0.039992707 0.656161307 0.000000000  0.842644636
#>    57   58 0.090447552 0.014549975 0.300012161 0.009831486 -1.057053994
#>    36   45 0.329562524 0.752190449 0.400603455 0.731749485  0.365848194
#>    43   20 0.296864447 0.887090815 0.243700563 0.956784860  0.295443977
#>    14   29 0.834796540 0.923016771 0.444049320 0.994817170  3.394873172
#>    16   31 0.793319502 0.816535072 0.751255589 0.648520506  0.794631573
#>    15   88 0.663433266 0.630241367 0.669130206 0.579326284  0.713446973
#>    44   50 0.398183391 0.609540808 0.531107659 0.600365866  0.469084567
#>    50   72 0.531107659 0.600365866 0.475643188 0.489848100  0.576443561
#>    32   28 0.220269124 0.660056151 0.350198677 0.573927215  0.305776705
#>    34   32 0.094636672 0.656126413 0.220269124 0.660056151  0.155524292
#>    95   72 0.434637373 0.449060452 0.475643188 0.489848100  0.426469600
#>    11   37 0.129306338 0.884683824 0.111088633 0.923168323  0.028337636
#>    69   14 1.000000000 0.791634646 0.834796540 0.923016771  0.923579822
#>    84   62 0.900223944 0.761698883 0.971229201 0.712553689  0.954037697
#>    56   63 0.815160589 0.449315232 0.855930571 0.540267194  0.795172851
#>    95   46 0.434637373 0.449060452 0.403639354 0.324366784  0.432668074
#>    22   70 0.674518335 0.458796311 0.670201896 0.364082324  0.741623797
#>     7   91 0.595836265 0.360040660 0.605395816 0.375299735  0.579762838
#>    60   90 0.488819005 0.241442017 0.491011202 0.353055140  0.460751281
#>    21   60 0.304782159 0.152926316 0.488819005 0.241442017  0.393292635
#>    85   69 0.985809812 0.374900239 1.000000000 0.791634646  1.496824728
#>    79    3 0.110285205 0.362230271 0.254231892 0.357893405  0.183579415
#>    96   80 0.549811220 0.232597348 0.421887029 0.055907624  0.506614290
#>    40   35 0.117655330 0.272817227 0.071465886 0.121255598  0.058995804
#>    21    5 0.304782159 0.152926316 0.347899572 0.061781561  0.391179663
#>    19   17 0.730271147 0.032520057 0.674517468 0.136634082  0.697617093
#>    51   85 0.927730467 0.039992707 0.985809812 0.374900239 54.293478428
#>    57   77 0.090447552 0.014549975 0.656161307 0.000000000 -1.057053994
#>           my1           mx2          my2 bp1 bp2
#>    0.91738800   0.521265945   0.92570368   0   0
#>    0.94299169   0.339560680   1.00035418   0   0
#>    0.90659988   0.458337133   0.83484583   0   0
#>    0.96958535   0.547817934   0.91738800   0   0
#>    0.86258182   0.710659861   0.83537903   0   0
#>    0.84800357   0.556863399   0.75299924   0   0
#>    0.83568854   0.362686923   0.89364524   0   0
#>    0.89364524   0.282428657   0.80707379   0   0
#>    0.73369347   0.495822594   0.69582413   0   0
#>    0.74822244   0.556863399   0.75299924   0   0
#>    0.79344408   0.433234009   0.83568854   0   0
#>    0.75648393   0.577696212   0.74153178   0   0
#>    0.74153178   0.630813770   0.67532119   0   0
#>    0.74462937   0.513700991   0.73369347   0   0
#>    0.83537903   0.778631098   0.90014205   0   0
#>    0.85238132   0.295443977   0.94113147   0   0
#>    0.92597607   0.282531066   0.85238132   0   0
#>    0.79159282   0.710659861   0.83537903   0   0
#>    0.94113147   0.212816377   0.89802094   0   0
#>    0.81801814   0.204450343   0.87373563   0   0
#>    0.67155269   0.354485932   0.70514074   0   0
#>    0.76613316   0.730257584   0.79159282   0   0
#>    0.67173567   0.705447878   0.64849293   0   0
#>   15.95443984   0.456294029   0.94299169   0   0
#>    2.07178727   0.521265945   0.92570368   0   0
#>    1.14153131   0.666616803   0.96958535   0   0
#>    0.86112558   0.730257584   0.79159282   0   0
#>    0.70514074   0.270064227   0.79762180   0   0
#>    0.80588623   0.265439638   0.81801814   0   0
#>    0.89802094   0.224672175   0.97541965   0   0
#>    0.96929796   0.345619819   0.92597607   0   0
#>    0.67532119   0.616533429   0.59921744   0   0
#>    0.57509408   0.495822594   0.69582413   0   0
#>    0.99858260   0.332368229   0.96929796   0   0
#>    0.72693377   0.798434050   0.64533341   0   0
#>    0.72721715   0.735869385   0.76613316   0   0
#>    0.61006119   0.746998153   0.57023990   0   0
#>    0.66926507   0.495822594   0.69582413   0   0
#>    0.79762180   0.234528573   0.73406393   0   0
#>    0.97541965   0.338751242   0.99858260   0   0
#>    0.50843699   0.539790141   0.46595599   0   0
#>    0.62609648   0.439598355   0.58206527   0   0
#>    0.88865078   0.164267602   0.92478789   0   0
#>    0.84572307   0.923579822   0.86509856   0   0
#>    0.73458258   0.904070936   0.69139004   0   0
#>    0.57471851   0.370531680   0.54332234   0   0
#>    0.65363150   0.386819703   0.57471851   0   0
#>    0.57023990   0.747810549   0.56500372   0   0
#>    0.52739973   0.426469600   0.49827831   0   0
#>    0.58206527   0.449668855   0.52739973   0   0
#>    0.64533341   0.846751890   0.61977544   0   0
#>    0.64798147   0.328251570   0.65363150   0   0
#>    0.71974819   0.234528573   0.73406393   0   0
#>    0.87373563   0.076208430   0.77582999   0   0
#>    0.51534815   0.745606200   0.52235962   0   0
#>    0.51467927   0.466104905   0.62609648   0   0
#>    0.86044167   0.050934672   0.92797663   0   0
#>    0.96619299   0.050934672   0.92797663   0   0
#>    0.61977544   0.878443221   0.63252252   0   0
#>    0.59247974   0.853159285   0.57403387   0   0
#>    0.56500372   0.842525881   0.59247974   0   0
#>    0.58277939   0.238957040   0.57041024   0   0
#>    0.68165208   0.878443221   0.63252252   0   0
#>    1.04227678   0.120612803   1.21355485   0   0
#>    0.76358230   1.496824728   0.56610851   0   0
#>    0.63252252   0.895604235   0.58947677   0   0
#>    0.51288863   0.852649190   0.56573278   0   0
#>    0.52235962   0.747810549   0.56500372   0   0
#>    0.48942225   0.370531680   0.54332234   0   0
#>    0.42479330   0.340923786   0.41042435   0   0
#>    0.54726114   0.346881865   0.42479330   0   0
#>    0.46595599   0.470282012   0.45423155   0   0
#>    0.56573278   0.853159285   0.57403387   0   0
#>    0.45423155   0.434765787   0.38458199   0   0
#>    0.38335022   0.460751281   0.29782139   0   0
#>    0.42912409   0.583417940   0.50993380   0   0
#>    0.57758663   0.139471674   0.56438946   0   0
#>    0.78939672  -0.176883035   0.73848600   0   0
#>    1.03075776  -0.002053485   0.98126765   0   0
#>    0.52140260   0.996374252   0.51278963   0   0
#>    0.50108431   0.889067323   0.38592414   0   0
#>    0.51757148   0.745606200   0.52235962   0   0
#>    0.51053241   0.889067323   0.38592414   0   0
#>    0.51278963   1.003416070   0.51053241   0   0
#>    0.98126765 -24.655715142   8.76552290   0   1
#>    1.21355485   0.064919288   5.69154585   0   1
#>    0.40828274   0.685552947   0.29871226   0   0
#>    0.38073437   0.625369779   0.42912409   0   0
#>    0.95206013  -0.002053485   0.98126765   0   0
#>    0.43544622   0.183579415   0.40390325   0   0
#>    0.41042435   0.343389963   0.40554417   0   0
#>    0.57041024   0.264446565   0.43544622   0   0
#>    0.34685240   0.635245649   0.32109318   0   0
#>    0.36234389   0.546811350   0.30571315   0   0
#>    0.40835611   0.868833627   0.33952402   0   0
#>    0.41267532   0.625369779   0.42912409   0   0
#>    0.26652298   0.338425164   0.38341912   0   0
#>    0.38978329   0.434765787   0.38458199   0   0
#>    0.32109318   0.566745597   0.29851407   0   0
#>    0.43274437   0.042087261   0.31159859   0   0
#>    0.29871226   0.715800335   0.23586468   0   0
#>    0.38341912   0.343389963   0.40554417   0   0
#>    0.56438946   0.057858165   0.58452813   0   0
#>    0.58452813   0.002338984   0.43142504   0   0
#>    0.29650171   0.503961080   0.13113935   0   0
#>    0.73848600  -2.459059152   0.70283514   0   1
#>    0.29851407   0.607116029   0.20651304   0   0
#>    0.20447769   0.430281406   0.26652298   0   0
#>    0.30571315   0.566745597   0.29851407   0   0
#>    0.55028964   1.496824728   0.56610851   0   0
#>    0.33952402   0.877915252   0.29659272   0   0
#>    0.21964236   0.607116029   0.20651304   0   0
#>    0.31713063   0.181204483   0.32507590   0   0
#>    0.45725755   0.002338984   0.43142504   0   0
#>    0.32507590   0.180015535   0.32296772   0   0
#>    0.20651304   0.536412981   0.12258219   0   0
#>    0.12921841   0.414494525  -0.06149282   0   0
#>    0.23830245   0.311202852   0.31713063   0   0
#>    0.25049711   0.180015535   0.32296772   0   0
#>    0.26813486   0.002338984   0.43142504   0   0
#>    0.22214602   0.715800335   0.23586468   0   0
#>    0.16404835   0.391179663   0.13802692   0   0
#>    0.06014159   0.584324547   0.05609704   0   0
#>    0.13823422   0.699881206   0.21964236   0   0
#>    0.21072174   0.324762301   0.23830245   0   0
#>    0.16388890   0.864863587   0.09201803   0   0
#>    0.12913924   0.839590456   0.16388890   0   0
#>    0.25987461   0.715800335   0.23586468   0   0
#>    0.08257348  -8.920689546  -1.53338344   0   1
#>    0.23656141   0.185938941   0.25049711   0   0
#>    0.20787503   0.163428169   0.08257348   0   0
#>    0.11325316   0.584324547   0.05609704   0   0
#>    0.12258219   0.545509797   0.11325316   0   0
#>    0.08380147   0.259336576   0.21072174   0   0
#>    0.13802692   0.382135631   0.02410930   0   0
#>    0.24262096   0.877915252   0.29659272   0   0
#>    0.08201884   0.654451859   0.13823422   0   0
#>   -0.06149282   0.410065257  -0.23389925   0   0
#>    0.03544891   0.729964808   0.09934124   0   0
#>    0.02410930   0.299558927   0.05829554   0   0
#>    0.05829554   0.410065257  -0.23389925   0   0
#>   -1.18630628  -1.057053994 -55.60611632   0   0
#>    0.08190438   0.196367184   0.06270342   0   0
#>    0.26138618   0.042087261   0.31159859   0   0
#>    0.08234313   0.229722836   0.08380147   0   0
#>   -9.04215453   1.011073307   0.24262096   0   0
#>    0.09387675   1.011073307   0.24262096   0   0
#>   -0.32427183   0.673960608   0.06014159   0   0
#>    0.06270342   0.229722836   0.08380147   0   0
#>    0.91738800   0.623704775   0.84800357   0   0
#>    0.94299169   0.486251856   0.90659988   0   0
#>    0.90659988   0.521265945   0.92570368   0   0
#>    0.96958535   0.654286543   0.86258182   0   0
#>    0.84800357   0.654286543   0.86258182   0   0
#>    0.83484583   0.433234009   0.83568854   0   0
#>    0.73369347   0.522254855   0.74822244   0   0
#>    0.74822244   0.458337133   0.83484583   0   0
#>    0.79344408   0.282428657   0.80707379   0   0
#>    0.75648393   0.674011972   0.67173567   0   0
#>    0.74462937   0.365804266   0.79344408   0   0
#>    1.00035418   0.457999886   5.77928709   0   1
#>    0.85238132   0.265439638   0.81801814   0   0
#>    0.92597607   0.362686923   0.89364524   0   0
#>    0.80707379   0.272904380   0.80588623   0   0
#>    0.75299924   0.577696212   0.74153178   0   0
#>    0.67155269   0.328251570   0.65363150   0   0
#>    0.76613316   0.717045025   0.75648393   0   0
#>   15.95443984   0.864590038   2.07178727   0   0
#>    2.07178727   0.736624341   1.14153131   0   0
#>    1.14153131   0.778631098   0.90014205   0   0
#>    0.90014205   0.836265562   0.86112558   0   0
#>    0.86112558   0.875807582   0.84572307   0   0
#>    0.80588623   0.270064227   0.79762180   0   0
#>    0.89802094   0.203945867   0.88865078   0   0
#>    0.67532119   0.674011972   0.67173567   0   0
#>    0.57509408   0.588762779   0.51534815   0   0
#>    0.99858260   0.339560680   1.00035418   0   0
#>    0.72693377   0.818798530   0.73458258   0   0
#>    0.72721715   0.705447878   0.64849293   0   0
#>    0.64849293   0.713446973   0.61006119   0   0
#>    0.66926507   0.353564070   0.67155269   0   0
#>    0.97541965   0.152548930   1.04227678   0   0
#>    0.59921744   0.597870200   0.57509408   0   0
#>    0.50843699   0.583417940   0.50993380   0   0
#>    0.88865078   0.204450343   0.87373563   0   0
#>    0.73458258   0.875807582   0.84572307   0   0
#>    0.57471851   0.439598355   0.58206527   0   0
#>    0.57023990   0.798434050   0.64533341   0   0
#>    0.52739973   0.564005287   0.51467927   0   0
#>    0.64798147   0.234528573   0.73406393   0   0
#>    0.71974819   0.076208430   0.77582999   0   0
#>    0.49827831   0.395652012   0.48942225   0   0
#>    0.54332234   0.314107857   0.54726114   0   0
#>    0.86044167  -0.013229307   0.78939672   0   0
#>    0.92478789   0.145757672   0.96619299   0   0
#>    0.59247974   0.846751890   0.61977544   0   0
#>    0.58277939   0.314107857   0.54726114   0   0
#>    0.68165208   1.130496060   0.55028964   0   0
#>    1.04227678   0.145757672   0.96619299   0   0
#>    0.86509856   0.954037697   0.76358230   0   0
#>    0.69139004   0.908677455   0.68165208   0   0
#>    0.58947677   0.853159285   0.57403387   0   0
#>    0.51288863   0.749121229   0.51757148   0   0
#>    0.48942225   0.346881865   0.42479330   0   0
#>    0.46595599   0.525279279   0.38978329   0   0
#>    0.38335022   0.434765787   0.38458199   0   0
#>    0.50993380   0.588762779   0.51534815   0   0
#>    0.57758663   0.238957040   0.57041024   0   0
#>    0.77582999  -0.013229307   0.78939672   0   0
#>    1.03075776   0.047541022   0.95206013   0   0
#>    0.52140260   0.852649190   0.56573278   0   0
#>    0.50108431   0.907844772   0.52140260   0   0
#>    0.51757148   0.741758723   0.40835611   0   0
#>    0.51053241   1.130496060   0.55028964   0   0
#>    0.51278963   0.895604235   0.58947677   0   0
#>    1.21355485   0.065126838   1.03075776   0   0
#>    0.40828274   0.741758723   0.40835611   0   0
#>    0.38073437   0.543037490   0.36234389   0   0
#>    0.95206013   0.050934672   0.92797663   0   0
#>    0.43544622   0.340923786   0.41042435   0   0
#>    0.34685240   0.645239090   0.41267532   0   0
#>    0.38978329   0.543037490   0.36234389   0   0
#>    0.32109318   0.685552947   0.29871226   0   0
#>    0.29782139   0.430281406   0.26652298   0   0
#>    0.43274437   0.183579415   0.40390325   0   0
#>    0.38341912   0.311202852   0.31713063   0   0
#>    0.56438946   0.126061661   0.45725755   0   0
#>    0.29650171   0.546811350   0.30571315   0   0
#>    0.73848600   0.057858165   0.58452813   0   0
#>    0.20447769   0.324762301   0.23830245   0   0
#>    0.33952402   0.889067323   0.38592414   0   0
#>    0.45725755   0.145780654   0.43274437   0   0
#>    0.12921841   0.536412981   0.12258219   0   0
#>    0.32296772   0.042087261   0.31159859   0   0
#>    0.26813486  -2.724196125  -0.12354999   0   1
#>    0.29659272   0.839771625   0.25987461   0   0
#>    0.22214602   0.699881206   0.21964236   0   0
#>    0.13113935   0.412737824   0.16404835   0   0
#>    0.13823422   0.545509797   0.11325316   0   0
#>    0.21072174   0.185938941   0.25049711   0   0
#>    0.16388890   0.839771625   0.25987461   0   0
#>    0.12913924   0.706675598   0.22214602   0   0
#>    0.08257348   0.165577637   0.08190438   0   0
#>    0.23656141   0.058995804   0.20787503   0   0
#>    0.09201803   0.829031365   0.03544891   0   0
#>    0.05609704   0.445203089  -1.18630628   0   0
#>    0.08201884   0.673960608   0.06014159   0   0
#>    0.03544891   0.842644636  -0.32427183   0   0
#>    0.02410930   0.414494525  -0.06149282   0   0
#>    0.09934124   0.732526253   0.12913924   0   0
#>    0.08190438   0.168589143   0.23656141   0   0
#>    0.26138618  -0.293620663   0.26813486   0   0
#>   -0.23389925   0.445203089  -1.18630628   0   0
#>    0.08234313   0.299558927   0.05829554   0   0
#>   -9.04215453   0.873522078   0.09387675   0   0
#>    0.09387675   0.864863587   0.09201803   0   0
#>   -0.32427183   1.373409315  -3.92841187   0   1
#>  -55.60611632   0.196367184   0.06270342   0   0
#>    0.74462937   0.354485932   0.70514074   0   0
#>    0.94113147   0.332368229   0.96929796   0   0
#>   15.95443984   3.849770511  18.43005103   0   1
#>    0.72693377   0.793499649   0.72721715   0   0
#>    0.61006119   0.616533429   0.59921744   0   0
#>    0.66926507   0.466104905   0.62609648   0   0
#>    0.50843699   0.564005287   0.51467927   0   0
#>    0.64798147   0.262554934   0.58277939   0   0
#>    0.71974819   0.159971054   0.57758663   0   0
#>    0.49827831   0.470282012   0.45423155   0   0
#>    0.86044167   0.164267602   0.92478789   0   0
#>    0.86509856   3.872445444   4.57308244   0   1
#>    0.76358230   0.904070936   0.69139004   0   0
#>    0.51288863   0.821506597   0.50108431   0   0
#>    0.38335022   0.343389963   0.40554417   0   0
#>    0.40828274   0.645239090   0.41267532   0   0
#>    0.38073437   0.633845673   0.34685240   0   0
#>    0.29782139   0.527940785   0.29650171   0   0
#>    0.20447769   0.412737824   0.16404835   0   0
#>    0.56610851   3.893655567   0.48449422   0   1
#>    0.40390325   0.181204483   0.32507590   0   0
#>    0.12921841   0.503961080   0.13113935   0   0
#>    0.20787503  -0.116590554   0.26138618   0   0
#>    0.13802692   0.273471189   0.08234313   0   0
#>    0.08201884   0.729964808   0.09934124   0   0
#>   -9.04215453  57.192200592  -9.54484816   0   1
#>  -55.60611632  -1.102488056 -57.37262620   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.601328297 0.916147676
#>   [2,] 0.249691779 0.283609264
#>   [3,] 0.254231892 0.357893405
#>   [4,] 0.278682859 0.889736188
#>   [5,] 0.347899572 0.061781561
#>   [6,] 0.900838945 0.165261818
#>   [7,] 0.595836265 0.360040660
#>   [8,] 0.626720269 0.266345355
#>   [9,] 0.402120197 0.752255743
#>  [10,] 0.000000000 0.960338310
#>  [11,] 0.129306338 0.884683824
#>  [12,] 0.818452135 0.602493701
#>  [13,] 0.006843302 0.522266934
#>  [14,] 0.834796540 0.923016771
#>  [15,] 0.663433266 0.630241367
#>  [16,] 0.793319502 0.816535072
#>  [17,] 0.674517468 0.136634082
#>  [18,] 0.868500737 0.594745014
#>  [19,] 0.730271147 0.032520057
#>  [20,] 0.243700563 0.956784860
#>  [21,] 0.304782159 0.152926316
#>  [22,] 0.674518335 0.458796311
#>  [23,] 0.346143712 0.515839324
#>  [24,] 0.911649256 0.611946998
#>  [25,] 0.011646144 0.997223031
#>  [26,] 0.710583225 0.694011888
#>  [27,] 0.195207658 0.798362838
#>  [28,] 0.350198677 0.573927215
#>  [29,] 0.444049320 0.994817170
#>  [30,] 0.141950325 0.180000735
#>  [31,] 0.751255589 0.648520506
#>  [32,] 0.220269124 0.660056151
#>  [33,] 0.302047796 0.504038676
#>  [34,] 0.094636672 0.656126413
#>  [35,] 0.071465886 0.121255598
#>  [36,] 0.329562524 0.752190449
#>  [37,] 0.111088633 0.923168323
#>  [38,] 0.878222938 0.568023455
#>  [39,] 0.583097386 0.957640422
#>  [40,] 0.117655330 0.272817227
#>  [41,] 0.790963119 0.126624382
#>  [42,] 0.942556562 0.126683447
#>  [43,] 0.296864447 0.887090815
#>  [44,] 0.398183391 0.609540808
#>  [45,] 0.400603455 0.731749485
#>  [46,] 0.403639354 0.324366784
#>  [47,] 0.017245808 0.989621770
#>  [48,] 0.204418673 0.485600281
#>  [49,] 0.662553387 0.787711691
#>  [50,] 0.531107659 0.600365866
#>  [51,] 0.927730467 0.039992707
#>  [52,] 0.553834020 0.864202436
#>  [53,] 0.407261598 0.544323536
#>  [54,] 0.778347585 0.165493016
#>  [55,] 0.670419076 0.714406241
#>  [56,] 0.815160589 0.449315232
#>  [57,] 0.090447552 0.014549975
#>  [58,] 0.300012161 0.009831486
#>  [59,] 0.314642350 0.871850793
#>  [60,] 0.488819005 0.241442017
#>  [61,] 0.090943273 0.973153295
#>  [62,] 0.971229201 0.712553689
#>  [63,] 0.855930571 0.540267194
#>  [64,] 0.827548309 0.571137371
#>  [65,] 0.407785633 0.921020138
#>  [66,] 0.483261480 0.558324254
#>  [67,] 0.057108232 0.504039557
#>  [68,] 0.584070790 0.645133186
#>  [69,] 1.000000000 0.791634646
#>  [70,] 0.670201896 0.364082324
#>  [71,] 0.747552257 0.324499577
#>  [72,] 0.475643188 0.489848100
#>  [73,] 0.487176380 0.986374795
#>  [74,] 0.642554159 0.154471025
#>  [75,] 0.234926034 1.000000000
#>  [76,] 0.845111679 0.652894085
#>  [77,] 0.656161307 0.000000000
#>  [78,] 0.200938254 0.963335115
#>  [79,] 0.110285205 0.362230271
#>  [80,] 0.421887029 0.055907624
#>  [81,] 0.344585822 0.844090935
#>  [82,] 0.572821313 0.406000955
#>  [83,] 0.650817257 0.118435767
#>  [84,] 0.900223944 0.761698883
#>  [85,] 0.985809812 0.374900239
#>  [86,] 0.380379296 0.306089180
#>  [87,] 0.193005886 0.179006573
#>  [88,] 0.669130206 0.579326284
#>  [89,] 0.501859345 0.017278378
#>  [90,] 0.491011202 0.353055140
#>  [91,] 0.605395816 0.375299735
#>  [92,] 0.331289682 0.021660458
#>  [93,] 0.718040065 0.902698714
#>  [94,] 0.869792074 0.481365748
#>  [95,] 0.434637373 0.449060452
#>  [96,] 0.549811220 0.232597348
#>  [97,] 0.487029549 0.422344244
#>  [98,] 0.778257492 0.839498251
#>  [99,] 0.081588609 0.386792806
#> [100,] 0.309897352 0.734239018
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,39,52), [2,7,5]
#> 2: (73,52,39), [1,30,4]
#> 3: (73,29,65), [17,4,29]
#> 4: (73,65,52), [9,2,3]
#> 5: (93,39,1), [1,6,31]
#> 6: (1,49,93), [18,5,7]
#> 7: (1,52,49), [25,6,1]
#> 8: (9,65,81), [10,13,9]
#> 9: (9,52,65), [4,8,12]
#> 10: (65,59,81), [22,8,20]
#> 11: (9,45,68), [40,12,16]
#> 12: (9,68,52), [25,9,11]
#> 13: (36,9,81), [8,22,16]
#> 14: (26,49,55), [15,28,27]
#> 15: (49,68,55), [38,14,25]
#> 16: (36,45,9), [11,13,34]
#> 17: (75,65,29), [3,0,41]
#> 18: (49,98,93), [32,6,21]
#> 19: (59,43,4), [23,24,20]
#> 20: (65,43,59), [19,10,37]
#> 21: (16,98,49), [18,27,33]
#> 22: (36,81,59), [10,35,13]
#> 23: (43,20,4), [36,19,37]
#> 24: (59,4,27), [64,35,19]
#> 25: (49,52,68), [12,15,7]
#> 26: (44,45,100), [34,57,46]
#> 27: (26,16,49), [21,14,43]
#> 28: (55,15,26), [44,14,38]
#> 29: (14,29,73), [3,30,0]
#> 30: (14,73,39), [2,31,29]
#> 31: (14,39,93), [5,32,30]
#> 32: (14,93,98), [18,33,31]
#> 33: (14,98,16), [21,54,32]
#> 34: (45,36,100), [47,26,16]
#> 35: (36,59,27), [24,47,22]
#> 36: (4,20,78), [48,53,23]
#> 37: (20,43,65), [20,41,23]
#> 38: (55,68,15), [49,28,15]
#> 39: (88,68,50), [40,65,49]
#> 40: (50,68,45), [11,46,39]
#> 41: (75,20,65), [37,17,48]
#> 42: (16,31,76), [61,55,43]
#> 43: (31,16,26), [27,44,42]
#> 44: (31,26,15), [28,45,43]
#> 45: (15,88,31), [58,44,49]
#> 46: (44,50,45), [40,26,51]
#> 47: (36,27,100), [52,34,35]
#> 48: (78,20,75), [41,77,36]
#> 49: (88,15,68), [38,39,45]
#> 50: (50,72,82), [89,94,66]
#> 51: (50,44,66), [60,66,46]
#> 52: (32,100,27), [47,63,62]
#> 53: (4,78,11), [70,64,36]
#> 54: (16,84,14), [78,33,55]
#> 55: (16,76,84), [81,54,42]
#> 56: (44,28,53), [68,60,57]
#> 57: (100,28,44), [56,26,62]
#> 58: (31,88,12), [74,61,45]
#> 59: (66,53,72), [67,66,60]
#> 60: (44,53,66), [59,51,56]
#> 61: (31,12,76), [72,42,58]
#> 62: (32,28,100), [57,52,75]
#> 63: (34,32,27), [52,96,95]
#> 64: (4,11,27), [96,24,53]
#> 65: (50,22,88), [84,39,94]
#> 66: (72,50,66), [51,59,50]
#> 67: (95,72,53), [59,86,91]
#> 68: (23,53,28), [56,88,86]
#> 69: (11,37,10), [98,97,70]
#> 70: (37,11,78), [53,71,69]
#> 71: (78,61,37), [98,70,77]
#> 72: (12,18,76), [80,61,73]
#> 73: (12,64,18), [85,72,74]
#> 74: (88,64,12), [73,58,84]
#> 75: (28,32,33), [112,88,62]
#> 76: (62,76,24), [80,136,81]
#> 77: (78,75,61), [106,71,48]
#> 78: (69,14,84), [54,79,0]
#> 79: (84,62,69), [135,78,81]
#> 80: (76,18,24), [82,76,72]
#> 81: (62,84,76), [55,76,79]
#> 82: (38,24,18), [80,85,104]
#> 83: (56,63,64), [90,102,101]
#> 84: (22,64,88), [74,65,102]
#> 85: (38,18,64), [73,90,82]
#> 86: (95,53,23), [68,87,67]
#> 87: (23,33,95), [111,86,88]
#> 88: (28,33,23), [87,68,75]
#> 89: (82,72,97), [91,118,50]
#> 90: (63,38,64), [85,83,100]
#> 91: (72,95,97), [120,89,67]
#> 92: (95,46,90), [122,120,123]
#> 93: (91,22,82), [94,108,116]
#> 94: (50,82,22), [93,65,50]
#> 95: (32,34,48), [127,112,63]
#> 96: (34,27,11), [64,97,63]
#> 97: (11,10,34), [131,96,69]
#> 98: (61,10,37), [69,71,109]
#> 99: (61,25,47), [105,109,106]
#> 100: (63,94,38), [104,90,101]
#> 101: (63,56,94), [121,100,83]
#> 102: (56,64,22), [84,115,83]
#> 103: (24,94,85), [121,136,104]
#> 104: (38,94,24), [103,82,100]
#> 105: (47,25,10), [0,109,99]
#> 106: (61,75,25), [0,99,77]
#> 107: (22,70,71), [125,115,116]
#> 108: (7,91,82), [93,114,113]
#> 109: (61,47,10), [105,98,99]
#> 110: (33,48,3), [141,111,112]
#> 111: (33,3,95), [123,87,110]
#> 112: (32,48,33), [110,75,95]
#> 113: (91,7,70), [119,116,108]
#> 114: (82,90,7), [134,108,118]
#> 115: (22,71,56), [137,102,107]
#> 116: (70,22,91), [93,113,107]
#> 117: (60,46,86), [126,133,122]
#> 118: (82,97,90), [120,114,89]
#> 119: (70,7,8), [132,125,113]
#> 120: (95,90,97), [118,91,92]
#> 121: (56,85,94), [103,101,137]
#> 122: (60,90,46), [92,117,129]
#> 123: (46,95,3), [111,126,92]
#> 124: (48,99,79), [149,141,142]
#> 125: (70,8,71), [138,107,119]
#> 126: (86,46,3), [123,140,117]
#> 127: (48,34,67), [128,142,95]
#> 128: (34,13,67), [130,127,131]
#> 129: (90,60,96), [153,134,122]
#> 130: (99,67,13), [128,150,142]
#> 131: (34,10,13), [0,128,97]
#> 132: (7,96,8), [144,119,134]
#> 133: (21,60,86), [117,146,154]
#> 134: (90,96,7), [132,114,129]
#> 135: (85,69,62), [79,136,0]
#> 136: (24,85,62), [135,76,103]
#> 137: (56,71,85), [151,121,115]
#> 138: (54,71,8), [125,152,160]
#> 139: (17,8,74), [144,156,152]
#> 140: (86,3,2), [143,146,126]
#> 141: (79,3,48), [110,124,143]
#> 142: (48,67,99), [130,124,127]
#> 143: (3,79,2), [148,140,141]
#> 144: (8,96,74), [165,139,132]
#> 145: (96,80,89), [172,165,153]
#> 146: (21,86,2), [140,157,133]
#> 147: (87,2,40), [148,162,157]
#> 148: (40,2,79), [143,149,147]
#> 149: (40,79,99), [124,179,148]
#> 150: (35,99,13), [130,0,179]
#> 151: (6,85,71), [137,160,168]
#> 152: (17,54,8), [138,139,159]
#> 153: (80,96,60), [129,154,145]
#> 154: (60,21,80), [167,153,133]
#> 155: (19,83,77), [170,184,171]
#> 156: (83,17,74), [139,164,171]
#> 157: (87,21,2), [146,147,166]
#> 158: (54,41,6), [169,160,159]
#> 159: (17,41,54), [158,152,176]
#> 160: (6,71,54), [138,158,151]
#> 161: (30,35,57), [0,178,163]
#> 162: (30,87,40), [147,163,178]
#> 163: (40,35,30), [161,162,179]
#> 164: (74,89,83), [170,156,165]
#> 165: (96,89,74), [164,144,145]
#> 166: (58,21,87), [157,186,181]
#> 167: (21,5,80), [174,154,181]
#> 168: (42,85,6), [151,183,182]
#> 169: (51,6,41), [158,173,183]
#> 170: (77,83,89), [164,177,155]
#> 171: (19,17,83), [156,155,176]
#> 172: (80,92,89), [180,145,174]
#> 173: (51,41,19), [176,184,169]
#> 174: (80,5,92), [175,172,167]
#> 175: (5,58,92), [180,174,181]
#> 176: (17,19,41), [173,159,171]
#> 177: (89,58,77), [185,170,180]
#> 178: (30,57,87), [186,162,161]
#> 179: (35,40,99), [149,150,163]
#> 180: (89,92,58), [175,177,172]
#> 181: (5,21,58), [166,175,167]
#> 182: (51,85,42), [168,183,0]
#> 183: (51,42,6), [168,169,182]
#> 184: (51,19,77), [155,0,173]
#> 185: (57,77,58), [177,186,0]
#> 186: (57,58,87), [166,178,185]
#> boundary nodes:  29 75 25 10 13 35 57 77 51 85 69 14 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.097913237 0.082225646 0.156711144 0.045321233 0.127674535 0.132931316 
#>                                                                         
#> 0.099560627 0.040831621 0.202885796 0.221146102 0.108369307 0.073726237 
#>                                                                         
#> 0.110701322 0.020562275 0.087295114 0.018373028 0.023416065 0.126764695 
#>                                                                         
#> 0.075625958 0.123762669 0.090740260 0.133905056 0.079308247 0.043945614 
#>                                                                         
#> 0.100132431 0.145698723 0.027462187 0.026626544 0.140232397 0.043261078 
#>                                                                         
#> 0.115994433 0.080747563 0.069348440 0.167937564 0.093957936 0.147841882 
#>                                                                         
#> 0.107389101 0.185183699 0.131398593 0.044096946 0.128350815 0.099304666 
#>                                                                         
#> 0.106382128 0.174081034 0.121966577 0.064284950 0.059756679 0.151108463 
#>                                                                         
#> 0.087427770 0.077278730 0.057016927 0.165299942 0.140558944 0.108601549 
#>                                                                         
#> 0.120650347 0.063692653 0.117142188 0.053891863 0.062676651 0.047269674 
#>                                                                         
#> 0.032649040 0.176151081 0.078127544 0.146464266 0.084151985 0.046451081 
#>                                                                         
#> 0.041934718 0.158629612 0.067429566 0.143536063 0.045647620 0.068457431 
#>                                                                         
#> 0.050770212 0.058810678 0.091961196 0.114584614 0.202808406 0.318592180 
#>                                                                         
#> 0.009441155 0.087066856 0.063339046 0.189838581 0.157464346 0.137125805 
#>                                                                         
#> 0.038679650 0.223297159 0.086890000 0.044762296 0.033984393 0.137078091 
#>                                                                         
#> 0.202132551 0.099355020 0.074475380 0.105057561 0.141950112 0.108395576 
#>                                                                         
#> 0.029582108 0.069403411 0.098654102 0.037773176 0.134098034 0.153122905 
#>                                                                         
#> 0.156648637 0.053467752 0.061630175 0.438124823 0.083987677 0.126247480 
#>                                                                         
#> 0.135499455 0.337968115 0.243530058 0.112989283 0.074422753 0.119775162 
#>                                                                         
#> 0.160048291 0.121264072 0.088813232 0.132606839 0.132476760 0.154725887 
#>                                                                         
#> 0.182104468 0.152072933 0.118556272 0.036603410 0.141820248 0.116471235 
#>                                                                         
#> 0.040864685 0.161961226 0.108380781 0.120325163 0.091755366 0.180059093 
#>                                                                         
#> 0.196511712 0.114778548 0.074220260 0.226204146 0.029881075 0.170625944 
#>                                                                         
#> 0.111978298 0.043423397 0.033439616 0.356284820 0.193814820 0.119545979 
#>                                                                         
#> 0.143174311 0.251957178 0.056821213 0.080930974 0.200176305 0.070384759 
#>                                                                         
#> 0.102830549 0.139173727 0.117484090 0.142282514 0.168859463 0.211142749 
#>                                                                         
#> 0.188544812 0.093120346 0.045045405 0.072557703 0.209187500 0.040161827 
#>                                                                         
#> 0.105324715 0.120586937 0.162751800 0.152787657 0.105292537 0.353346876 
#>                                                                         
#> 0.254069402 0.118511172 0.100856369 0.114274656 0.142067306 0.107056180 
#>                                                                         
#> 0.084454293 0.139616931 0.190048045 0.171641493 0.061022185 0.089704445 
#>                                                                         
#> 0.122232636 0.049994829 0.107543724 0.198790712 0.149461939 0.120148106 
#>                                                                         
#> 0.065846081 0.081448404 0.068898636 0.116345679 0.174200240 0.099118557 
#>                                                                         
#> 0.058229252 0.149812317 0.098419133 0.050644888 0.084869990 0.116925088 
#>                                                                         
#> 0.110432299 0.104170107 0.139516658 0.028435241 0.122450354 0.110862734 
#>                                                                         
#> 0.087334585 0.111332963 0.201515658 0.175174448 0.231171961 0.075515078 
#>                                                                         
#> 0.035599995 0.060510508 0.140961465 0.248376637 0.055195981 0.082869698 
#>                                                                         
#> 0.152871078 0.051400747 0.091841727 0.153768660 0.065769737 0.097448210 
#>                                                                         
#> 0.106972739 0.118878467 0.155181407 0.136370251 0.148460004 0.134042984 
#>                                                                         
#> 0.160081341 0.170803366 0.186168602 0.157639278 0.220593838 0.089716282 
#>                                                                         
#> 0.406184896 0.221028217 0.138237429 0.197238182 0.036970510 0.118974844 
#>                                                                         
#> 0.122491577 0.107766096 0.173281533 0.095943462 0.161896123 0.155266344 
#>                                                                         
#> 0.117023202 0.197600667 0.096854260 0.116875077 0.051065239 0.265730085 
#>                                                                         
#> 0.201984510 0.070654190 0.087949402 0.128123004 0.274498134 0.209617722 
#>                                                                         
#> 0.073923251 0.087656480 0.397289174 0.173200079 0.051232810 0.133240537 
#>                                                                         
#> 0.123654697 0.155884195 0.125693898 0.057836918 0.042578651 0.211076873 
#>                                                                         
#> 0.086353903 0.099671715 0.128488864 0.094812294 0.018006232 0.111634649 
#>                                                                         
#> 0.204217017 0.416975931 0.144012003 0.218137245 0.158443656 0.100828952 
#>                                     
#> 0.118102510 0.339906259 0.565900834 
#> 
#> $alpha
#> [1] 0.2745146
#> 
#> attr(,"class")
#> [1] "scree"