Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1          y1         x2          y2         mx1
#>    52   18 0.84708485 0.835838023 0.83919312 0.960526048  0.81477298
#>    18   25 0.83919312 0.960526048 0.69828308 0.961055330  0.76859317
#>    25   24 0.69828308 0.961055330 0.64101545 0.856550656  0.71103589
#>    24   72 0.64101545 0.856550656 0.64322461 0.712534895  0.69992280
#>    72   70 0.64322461 0.712534895 0.78567784 0.715806362  0.71311835
#>    70   52 0.78567784 0.715806362 0.84708485 0.835838023  0.80524036
#>    19   31 0.53319387 0.825880930 0.57444788 0.869949466  0.52537424
#>    31   67 0.57444788 0.869949466 0.56585112 0.970495930  0.52832677
#>    67   35 0.56585112 0.970495930 0.46598235 1.000000000  0.50285770
#>    35    8 0.46598235 1.000000000 0.43163733 0.994097625  0.45905536
#>     8   32 0.43163733 0.994097625 0.35753106 0.897964607  0.42332928
#>    32   46 0.35753106 0.897964607 0.41802574 0.839673448  0.41936270
#>    46   16 0.41802574 0.839673448 0.50905163 0.777485326  0.47145069
#>    16   19 0.50905163 0.777485326 0.53319387 0.825880930  0.47904231
#>    90   22 0.26629962 0.157981594 0.33765631 0.052308369  0.32323382
#>    22   28 0.33765631 0.052308369 0.49321211 0.020933667  0.42181182
#>    28   83 0.49321211 0.020933667 0.52549889 0.106055785  0.44494030
#>    83   82 0.52549889 0.106055785 0.49357257 0.203470477  0.45954316
#>    82   78 0.49357257 0.203470477 0.40028742 0.273560897  0.42303723
#>    78   37 0.40028742 0.273560897 0.34497874 0.232067155  0.40581937
#>    37   90 0.34497874 0.232067155 0.26629962 0.157981594  0.32824707
#>    76   71 0.81950737 0.550338442 0.93919269 0.515509270  0.89821837
#>    71   98 0.93919269 0.515509270 0.97262248 0.518046031  0.94935157
#>    98   38 0.97262248 0.518046031 0.99673580 0.724579585  1.00207033
#>    38   30 0.99673580 0.724579585 0.88313867 0.703359952  0.93108969
#>    30   76 0.88313867 0.703359952 0.81950737 0.550338442  0.86904246
#>    78   65 0.40028742 0.273560897 0.52433176 0.263726135  0.46728257
#>    65   26 0.52433176 0.263726135 0.59125601 0.248926997  0.56256676
#>    26   66 0.59125601 0.248926997 0.66339389 0.300800737  0.60788985
#>    66   42 0.66339389 0.300800737 0.54014426 0.412678123  0.61964440
#>    42   78 0.54014426 0.412678123 0.40028742 0.273560897  0.46134389
#>    50   85 0.03137898 0.025183476 0.15580548 0.201977054 -0.04724438
#>    85   79 0.15580548 0.201977054 0.19777873 0.341699119  0.13893835
#>    79   95 0.19777873 0.341699119 0.13659043 0.436231728  0.15191553
#>    95   62 0.13659043 0.436231728 0.09523562 0.476262652  0.07630353
#>    62    7 0.09523562 0.476262652 0.03362977 0.504494081  0.02959279
#>     7   50 0.03362977 0.504494081 0.03137898 0.025183476 -0.31456706
#>    88   93 0.02619324 0.759321770 0.00000000 0.999363105 -3.08840640
#>    93   50 0.00000000 0.999363105 0.03137898 0.025183476 -6.68652174
#>    62   48 0.09523562 0.476262652 0.12933585 0.553177557  0.08178443
#>    48   44 0.12933585 0.553177557 0.08148396 0.669676058  0.05086688
#>    44   88 0.08148396 0.669676058 0.02619324 0.759321770 -0.08553273
#>    35   93 0.46598235 1.000000000 0.00000000 0.999363105  0.23126462
#>    93   29 0.00000000 0.999363105 0.13408639 0.979741594  0.22925336
#>    29   45 0.13408639 0.979741594 0.18248466 0.918311515  0.28006115
#>    45   53 0.18248466 0.918311515 0.23812439 0.887050148  0.28363957
#>    53   32 0.23812439 0.887050148 0.35753106 0.897964607  0.28523001
#>    40   33 1.00000000 0.178298366 0.98767239 0.459659903  1.22548067
#>    33   71 0.98767239 0.459659903 0.93919269 0.515509270  0.94077432
#>    71   69 0.93919269 0.515509270 0.84567710 0.445948479  0.91194502
#>    69   27 0.84567710 0.445948479 0.85208795 0.300528554  0.90597645
#>    27   40 0.85208795 0.300528554 1.00000000 0.178298366  0.98154292
#>    69   43 0.84567710 0.445948479 0.75758024 0.379060885  0.81617625
#>    43   80 0.75758024 0.379060885 0.75443116 0.359452319  0.77776196
#>    80   23 0.75443116 0.359452319 0.74309293 0.332491397  0.76287202
#>    23   27 0.74309293 0.332491397 0.85208795 0.300528554  0.78170882
#>    55   20 0.63917208 0.193672214 0.76542002 0.002412271  0.73465468
#>    20   68 0.76542002 0.002412271 0.80989626 0.004717148  0.78522817
#>    68   15 0.80989626 0.004717148 0.90674226 0.000000000  0.85872891
#>    15   96 0.90674226 0.000000000 0.91999591 0.030582210  0.87528654
#>    96   73 0.91999591 0.030582210 0.88073431 0.181852425  0.89684123
#>    73   55 0.88073431 0.181852425 0.63917208 0.193672214  0.75825469
#>    92   34 0.97743970 0.934522087 0.92581855 0.909695656  0.95915456
#>    34   52 0.92581855 0.909695656 0.84708485 0.835838023  0.87851532
#>    52   21 0.84708485 0.835838023 0.93344216 0.866502335  0.88140201
#>    21   92 0.93344216 0.866502335 0.97743970 0.934522087  0.99473577
#>    88   94 0.02619324 0.759321770 0.23137553 0.773540337  0.12792901
#>    94   45 0.23137553 0.773540337 0.18248466 0.918311515  0.14066325
#>    25   77 0.69828308 0.961055330 0.68732142 0.972680779  0.66703667
#>    77   97 0.68732142 0.972680779 0.57858968 0.985715566  0.63508920
#>    97   67 0.57858968 0.985715566 0.56585112 0.970495930  0.59876944
#>    31   24 0.57444788 0.869949466 0.64101545 0.856550656  0.61419959
#>    40   96 1.00000000 0.178298366 0.91999591 0.030582210  1.97042055
#>    46   61 0.41802574 0.839673448 0.38757072 0.791380765  0.44856638
#>    61   58 0.38757072 0.791380765 0.45266903 0.642344555  0.44165833
#>    58   57 0.45266903 0.642344555 0.61211685 0.634236405  0.53488888
#>    57   72 0.61211685 0.634236405 0.64322461 0.712534895  0.55537108
#>    72   19 0.64322461 0.712534895 0.53319387 0.825880930  0.58753490
#>    89   91 0.38413270 0.549202004 0.34177814 0.627881960  0.29581626
#>    91   56 0.34177814 0.627881960 0.32464736 0.710653799  0.33343610
#>    56   64 0.32464736 0.710653799 0.23471943 0.753390002  0.26078688
#>    64   44 0.23471943 0.753390002 0.08148396 0.669676058  0.19139589
#>    48   89 0.12933585 0.553177557 0.38413270 0.549202004  0.25620059
#>    86   77 0.98171725 0.976294476 0.68732142 0.972680779  0.82512692
#>    34   86 0.92581855 0.909695656 0.98171725 0.976294476  0.90912108
#>    55   83 0.63917208 0.193672214 0.52549889 0.106055785  0.64926128
#>    28   81 0.49321211 0.020933667 0.14667600 0.018212048  0.35201411
#>    52   30 0.84708485 0.835838023 0.88313867 0.703359952  0.91445953
#>    38   92 0.99673580 0.724579585 0.97743970 0.934522087  1.05543480
#>    90   74 0.26629962 0.157981594 0.22170483 0.139540015  0.27075718
#>    74   81 0.22170483 0.139540015 0.14667600 0.018212048  0.24060804
#>    80   60 0.75443116 0.359452319 0.74908039 0.378469791  0.71720946
#>    60   66 0.74908039 0.378469791 0.66339389 0.300800737  0.68461883
#>    66   55 0.66339389 0.300800737 0.63917208 0.193672214  0.73870202
#>    73   27 0.88073431 0.181852425 0.85208795 0.300528554  0.76857777
#>    65   47 0.52433176 0.263726135 0.55510371 0.171495191  0.54999044
#>    47   55 0.55510371 0.171495191 0.63917208 0.193672214  0.59219142
#>    73   40 0.88073431 0.181852425 1.00000000 0.178298366  0.94273737
#>    70   76 0.78567784 0.715806362 0.81950737 0.550338442  0.82522832
#>    53   49 0.23812439 0.887050148 0.27895331 0.869876810  0.29186844
#>    49  100 0.27895331 0.869876810 0.27675581 0.844865352  0.32934965
#>   100   61 0.27675581 0.844865352 0.38757072 0.791380765  0.34063601
#>    40   38 1.00000000 0.178298366 0.99673580 0.724579585  4.49375471
#>    92   39 0.97743970 0.934522087 0.97793459 0.949897176  0.94140023
#>    39   86 0.97793459 0.949897176 0.98171725 0.976294476  0.91952435
#>    97   86 0.57858968 0.985715566 0.98171725 0.976294476  0.83872902
#>    76   99 0.81950737 0.550338442 0.74541082 0.450406774  0.76377857
#>    99   69 0.74541082 0.450406774 0.84567710 0.445948479  0.79482010
#>    78   59 0.40028742 0.273560897 0.39722039 0.275370898  0.38010132
#>    59   75 0.39722039 0.275370898 0.26590234 0.256053671  0.32517575
#>    75   90 0.26590234 0.256053671 0.26629962 0.157981594  0.29423880
#>    86   92 0.98171725 0.976294476 0.97743970 0.934522087  1.63083619
#>    42   54 0.54014426 0.412678123 0.51949539 0.520968688  0.51072983
#>    54   87 0.51949539 0.520968688 0.44867997 0.555179850  0.47096005
#>    87   63 0.44867997 0.555179850 0.43193173 0.544956534  0.46437589
#>    63   89 0.43193173 0.544956534 0.38413270 0.549202004  0.40341576
#>    89   79 0.38413270 0.549202004 0.19777873 0.341699119  0.31518832
#>    79   59 0.19777873 0.341699119 0.39722039 0.275370898  0.32740618
#>    60   99 0.74908039 0.378469791 0.74541082 0.450406774  0.64686253
#>    99   84 0.74541082 0.450406774 0.66700789 0.565248465  0.63739156
#>    84   54 0.66700789 0.565248465 0.51949539 0.520968688  0.61147679
#>    64   94 0.23471943 0.753390002 0.23137553 0.773540337  0.13831867
#>    94   53 0.23137553 0.773540337 0.23812439 0.887050148  0.17173935
#>    65   82 0.52433176 0.263726135 0.49357257 0.203470477  0.54005239
#>    95   89 0.13659043 0.436231728 0.38413270 0.549202004  0.25597201
#>    53  100 0.23812439 0.887050148 0.27675581 0.844865352  0.25067159
#>    81   85 0.14667600 0.018212048 0.15580548 0.201977054  0.09454271
#>    56   61 0.32464736 0.710653799 0.38757072 0.791380765  0.31686731
#>    84   57 0.66700789 0.565248465 0.61211685 0.634236405  0.58845735
#>    58   87 0.45266903 0.642344555 0.44867997 0.555179850  0.51204929
#>    91   58 0.34177814 0.627881960 0.45266903 0.642344555  0.39120375
#>    84   72 0.66700789 0.565248465 0.64322461 0.712534895  0.69733362
#>    89   87 0.38413270 0.549202004 0.44867997 0.555179850  0.41192907
#>    79   75 0.19777873 0.341699119 0.26590234 0.256053671  0.31674802
#>    84   76 0.66700789 0.565248465 0.81950737 0.550338442  0.74955761
#>    90   85 0.26629962 0.157981594 0.15580548 0.201977054  0.22088270
#>     1   18 0.77292205 0.841992785 0.83919312 0.960526048  0.81477298
#>     1   25 0.77292205 0.841992785 0.69828308 0.961055330  0.76859317
#>     1   24 0.77292205 0.841992785 0.64101545 0.856550656  0.71103589
#>     1   72 0.77292205 0.841992785 0.64322461 0.712534895  0.69992280
#>     1   70 0.77292205 0.841992785 0.78567784 0.715806362  0.71311835
#>     1   52 0.77292205 0.841992785 0.84708485 0.835838023  0.80524036
#>     2   31 0.47613952 0.876843679 0.57444788 0.869949466  0.52537424
#>     2   67 0.47613952 0.876843679 0.56585112 0.970495930  0.52832677
#>     2   35 0.47613952 0.876843679 0.46598235 1.000000000  0.50285770
#>     2    8 0.47613952 0.876843679 0.43163733 0.994097625  0.45905536
#>     2   32 0.47613952 0.876843679 0.35753106 0.897964607  0.42332928
#>     2   46 0.47613952 0.876843679 0.41802574 0.839673448  0.41936270
#>     2   16 0.47613952 0.876843679 0.50905163 0.777485326  0.47145069
#>     2   19 0.47613952 0.876843679 0.53319387 0.825880930  0.47904231
#>     3   22 0.38651946 0.146282588 0.33765631 0.052308369  0.32323382
#>     3   28 0.38651946 0.146282588 0.49321211 0.020933667  0.42181182
#>     3   83 0.38651946 0.146282588 0.52549889 0.106055785  0.44494030
#>     3   82 0.38651946 0.146282588 0.49357257 0.203470477  0.45954316
#>     3   78 0.38651946 0.146282588 0.40028742 0.273560897  0.42303723
#>     3   37 0.38651946 0.146282588 0.34497874 0.232067155  0.40581937
#>     3   90 0.38651946 0.146282588 0.26629962 0.157981594  0.32824707
#>     4   71 0.92106768 0.686769649 0.93919269 0.515509270  0.89821837
#>     4   98 0.92106768 0.686769649 0.97262248 0.518046031  0.94935157
#>     4   38 0.92106768 0.686769649 0.99673580 0.724579585  1.00207033
#>     4   30 0.92106768 0.686769649 0.88313867 0.703359952  0.93108969
#>     4   76 0.92106768 0.686769649 0.81950737 0.550338442  0.86904246
#>     5   65 0.55474373 0.317934130 0.52433176 0.263726135  0.46728257
#>     5   26 0.55474373 0.317934130 0.59125601 0.248926997  0.56256676
#>     5   66 0.55474373 0.317934130 0.66339389 0.300800737  0.60788985
#>     5   42 0.55474373 0.317934130 0.54014426 0.412678123  0.61964440
#>     5   78 0.55474373 0.317934130 0.40028742 0.273560897  0.46134389
#>     6   85 0.09776455 0.355236889 0.15580548 0.201977054 -0.04724438
#>     6   79 0.09776455 0.355236889 0.19777873 0.341699119  0.13893835
#>     6   95 0.09776455 0.355236889 0.13659043 0.436231728  0.15191553
#>     6   62 0.09776455 0.355236889 0.09523562 0.476262652  0.07630353
#>     6    7 0.09776455 0.355236889 0.03362977 0.504494081  0.02959279
#>     6   50 0.09776455 0.355236889 0.03137898 0.025183476 -0.31456706
#>     7   93 0.03362977 0.504494081 0.00000000 0.999363105 -3.08840640
#>     7   48 0.03362977 0.504494081 0.12933585 0.553177557  0.08178443
#>     7   44 0.03362977 0.504494081 0.08148396 0.669676058  0.05086688
#>     7   88 0.03362977 0.504494081 0.02619324 0.759321770 -0.08553273
#>     8   93 0.43163733 0.994097625 0.00000000 0.999363105  0.23126462
#>     8   29 0.43163733 0.994097625 0.13408639 0.979741594  0.22925336
#>     8   45 0.43163733 0.994097625 0.18248466 0.918311515  0.28006115
#>     8   53 0.43163733 0.994097625 0.23812439 0.887050148  0.28363957
#>     9   33 0.97616690 0.436057792 0.98767239 0.459659903  1.22548067
#>     9   71 0.97616690 0.436057792 0.93919269 0.515509270  0.94077432
#>     9   69 0.97616690 0.436057792 0.84567710 0.445948479  0.91194502
#>     9   27 0.97616690 0.436057792 0.85208795 0.300528554  0.90597645
#>     9   40 0.97616690 0.436057792 1.00000000 0.178298366  0.98154292
#>    10   43 0.78391097 0.342388791 0.75758024 0.379060885  0.81617625
#>    10   80 0.78391097 0.342388791 0.75443116 0.359452319  0.77776196
#>    10   23 0.78391097 0.342388791 0.74309293 0.332491397  0.76287202
#>    10   27 0.78391097 0.342388791 0.85208795 0.300528554  0.78170882
#>    10   69 0.78391097 0.342388791 0.84567710 0.445948479  0.84981726
#>    11   20 0.83700197 0.054918631 0.76542002 0.002412271  0.73465468
#>    11   68 0.83700197 0.054918631 0.80989626 0.004717148  0.78522817
#>    11   15 0.83700197 0.054918631 0.90674226 0.000000000  0.85872891
#>    11   96 0.83700197 0.054918631 0.91999591 0.030582210  0.87528654
#>    11   73 0.83700197 0.054918631 0.88073431 0.181852425  0.89684123
#>    11   55 0.83700197 0.054918631 0.63917208 0.193672214  0.75825469
#>    12   34 0.93360865 0.884801355 0.92581855 0.909695656  0.95915456
#>    12   52 0.93360865 0.884801355 0.84708485 0.835838023  0.87851532
#>    12   21 0.93360865 0.884801355 0.93344216 0.866502335  0.88140201
#>    12   92 0.93360865 0.884801355 0.97743970 0.934522087  0.99473577
#>    13   88 0.03721209 0.828764877 0.02619324 0.759321770 -0.29440361
#>    13   94 0.03721209 0.828764877 0.23137553 0.773540337  0.12792901
#>    13   45 0.03721209 0.828764877 0.18248466 0.918311515  0.14066325
#>    13   29 0.03721209 0.828764877 0.13408639 0.979741594  0.09436123
#>    13   93 0.03721209 0.828764877 0.00000000 0.999363105  0.05722946
#>    14   77 0.63086086 0.939537020 0.68732142 0.972680779  0.66703667
#>    14   97 0.63086086 0.939537020 0.57858968 0.985715566  0.63508920
#>    14   67 0.63086086 0.939537020 0.56585112 0.970495930  0.59876944
#>    14   31 0.63086086 0.939537020 0.57444788 0.869949466  0.58228045
#>    14   24 0.63086086 0.939537020 0.64101545 0.856550656  0.61419959
#>    14   25 0.63086086 0.939537020 0.69828308 0.961055330  0.67954567
#>    15   20 0.90674226 0.000000000 0.76542002 0.002412271  0.82411370
#>    16   61 0.50905163 0.777485326 0.38757072 0.791380765  0.44856638
#>    16   58 0.50905163 0.777485326 0.45266903 0.642344555  0.44165833
#>    16   57 0.50905163 0.777485326 0.61211685 0.634236405  0.53488888
#>    16   72 0.50905163 0.777485326 0.64322461 0.712534895  0.55537108
#>    17   91 0.30249758 0.640521399 0.34177814 0.627881960  0.29581626
#>    17   56 0.30249758 0.640521399 0.32464736 0.710653799  0.33343610
#>    17   64 0.30249758 0.640521399 0.23471943 0.753390002  0.26078688
#>    17   44 0.30249758 0.640521399 0.08148396 0.669676058  0.19139589
#>    17   48 0.30249758 0.640521399 0.12933585 0.553177557  0.19085984
#>    17   89 0.30249758 0.640521399 0.38413270 0.549202004  0.25620059
#>    18   77 0.83919312 0.960526048 0.68732142 0.972680779  0.82512692
#>    18   34 0.83919312 0.960526048 0.92581855 0.909695656  0.86151922
#>    18   86 0.83919312 0.960526048 0.98171725 0.976294476  0.90912108
#>    19   24 0.53319387 0.825880930 0.64101545 0.856550656  0.60339421
#>    20   83 0.76542002 0.002412271 0.52549889 0.106055785  0.64926128
#>    20   28 0.76542002 0.002412271 0.49321211 0.020933667  0.62973532
#>    20   81 0.76542002 0.002412271 0.14667600 0.018212048  0.35201411
#>    21   30 0.93344216 0.866502335 0.88313867 0.703359952  0.91445953
#>    21   38 0.93344216 0.866502335 0.99673580 0.724579585  0.92780588
#>    22   74 0.33765631 0.052308369 0.22170483 0.139540015  0.27075718
#>    22   81 0.33765631 0.052308369 0.14667600 0.018212048  0.24060804
#>    23   60 0.74309293 0.332491397 0.74908039 0.378469791  0.71720946
#>    23   66 0.74309293 0.332491397 0.66339389 0.300800737  0.68461883
#>    23   55 0.74309293 0.332491397 0.63917208 0.193672214  0.73870202
#>    23   73 0.74309293 0.332491397 0.88073431 0.181852425  0.76107659
#>    26   47 0.59125601 0.248926997 0.55510371 0.171495191  0.54999044
#>    26   55 0.59125601 0.248926997 0.63917208 0.193672214  0.59219142
#>    30   70 0.88313867 0.703359952 0.78567784 0.715806362  0.84124312
#>    32   49 0.35753106 0.897964607 0.27895331 0.869876810  0.29186844
#>    32  100 0.35753106 0.897964607 0.27675581 0.844865352  0.32934965
#>    32   61 0.35753106 0.897964607 0.38757072 0.791380765  0.34063601
#>    33   38 0.98767239 0.459659903 0.99673580 0.724579585  4.49375471
#>    33   98 0.98767239 0.459659903 0.97262248 0.518046031  1.33524129
#>    34   39 0.92581855 0.909695656 0.97793459 0.949897176  0.94140023
#>    35   97 0.46598235 1.000000000 0.57858968 0.985715566  0.52538711
#>    35   86 0.46598235 1.000000000 0.98171725 0.976294476  0.83872902
#>    36   71 0.81970856 0.478421328 0.93919269 0.515509270  0.87880114
#>    36   76 0.81970856 0.478421328 0.81950737 0.550338442  0.87399790
#>    36   99 0.81970856 0.478421328 0.74541082 0.450406774  0.76377857
#>    36   69 0.81970856 0.478421328 0.84567710 0.445948479  0.79482010
#>    37   59 0.34497874 0.232067155 0.39722039 0.275370898  0.38010132
#>    37   75 0.34497874 0.232067155 0.26590234 0.256053671  0.32517575
#>    41   42 0.45231085 0.464974148 0.54014426 0.412678123  0.45084096
#>    41   54 0.45231085 0.464974148 0.51949539 0.520968688  0.51072983
#>    41   87 0.45231085 0.464974148 0.44867997 0.555179850  0.47096005
#>    41   63 0.45231085 0.464974148 0.43193173 0.544956534  0.46437589
#>    41   89 0.45231085 0.464974148 0.38413270 0.549202004  0.40341576
#>    41   79 0.45231085 0.464974148 0.19777873 0.341699119  0.31518832
#>    41   59 0.45231085 0.464974148 0.39722039 0.275370898  0.32740618
#>    41   78 0.45231085 0.464974148 0.40028742 0.273560897  0.45077515
#>    42   60 0.54014426 0.412678123 0.74908039 0.378469791  0.64628303
#>    42   99 0.54014426 0.412678123 0.74541082 0.450406774  0.64686253
#>    42   84 0.54014426 0.412678123 0.66700789 0.565248465  0.63739156
#>    43   99 0.75758024 0.379060885 0.74541082 0.450406774  0.79438216
#>    43   60 0.75758024 0.379060885 0.74908039 0.378469791  0.75083682
#>    44   94 0.08148396 0.669676058 0.23137553 0.773540337  0.13831867
#>    47   82 0.55510371 0.171495191 0.49357257 0.203470477  0.54005239
#>    47   83 0.55510371 0.171495191 0.52549889 0.106055785  0.50688327
#>    48   95 0.12933585 0.553177557 0.13659043 0.436231728  0.15442827
#>    51   94 0.24590523 0.789281125 0.23137553 0.773540337  0.18230301
#>    51   64 0.24590523 0.789281125 0.23471943 0.753390002  0.25426692
#>    51   56 0.24590523 0.789281125 0.32464736 0.710653799  0.29087720
#>    51   61 0.24590523 0.789281125 0.38757072 0.791380765  0.31686731
#>    51  100 0.24590523 0.789281125 0.27675581 0.844865352  0.31679790
#>    51   53 0.24590523 0.789281125 0.23812439 0.887050148  0.22567131
#>    54   57 0.51949539 0.520968688 0.61211685 0.634236405  0.58845735
#>    54   58 0.51949539 0.520968688 0.45266903 0.642344555  0.53076401
#>    56   58 0.32464736 0.710653799 0.45266903 0.642344555  0.39120375
#>    58   89 0.45266903 0.642344555 0.38413270 0.549202004  0.40065097
#>    70   84 0.78567784 0.715806362 0.66700789 0.565248465  0.71595441
#>    74   85 0.22170483 0.139540015 0.15580548 0.201977054  0.22088270
#>    75   85 0.26590234 0.256053671 0.15580548 0.201977054  0.19207564
#>    15   40 0.90674226 0.000000000 1.00000000 0.178298366  1.97042055
#>    38   86 0.99673580 0.724579585 0.98171725 0.976294476  1.63083619
#>    50   81 0.03137898 0.025183476 0.14667600 0.018212048  0.09454271
#>          my1         mx2         my2 bp1 bp2
#>   0.89638670  0.86151922  0.89934535   0   0
#>   0.92220549  0.76903091  1.03874419   0   0
#>   0.88612349  0.67954567  0.90337985   0   0
#>   0.78542945  0.60339421  0.78394873   0   0
#>   0.77220948  0.71595441  0.64871561   0   0
#>   0.78152181  0.84124312  0.76310315   0   0
#>   0.87454503  0.59658682  0.80788058   0   0
#>   0.91664684  0.58228045  0.92125990   0   0
#>   0.94104424  0.52538711  1.01730439   0   0
#>   0.93743169  0.23126462  2.26291117   0   0
#>   0.92387233  0.28523001  1.03032921   0   0
#>   0.90159728  0.36151435  0.84156214   0   0
#>   0.82016033  0.44856638  0.78666414   0   0
#>   0.82267503  0.58753490  0.76855330   0   0
#>   0.11949817  0.27075718  0.08406289   0   0
#>   0.06824122  0.32339412 -0.41971384   0   0
#>   0.08792738  0.62973532  0.01783472   0   0
#>   0.13837876  0.50688327  0.15389382   0   0
#>   0.20671619  0.46145347  0.25784535   0   0
#>   0.20857869  0.38010132  0.24285931   0   0
#>   0.17101471  0.29423880  0.20713161   0   0
#>   0.59776213  0.87399790  0.51453204   0   0
#>   0.60317372  0.95845099  0.48326038   0   0
#>   0.61928234  1.33524129  0.58038378   0   0
#>   0.76133422  0.92780588  0.77891372   0   0
#>   0.61948088  0.82522832  0.63770022   0   0
#>   0.33136699  0.46145347  0.25784535   0   0
#>   0.27791031  0.54999044  0.22103805   0   0
#>   0.30189116  0.64639721  0.24834115   0   0
#>   0.37643177  0.64628303  0.40577819   0   0
#>   0.35203863  0.45084096  0.36259740   0   0
#>   0.21270042  0.09454271  0.11291132   0   0
#>   0.28320955  0.19207564  0.26724683   0   0
#>   0.37908220  0.27566570  0.45918220   0   0
#>   0.41532775  0.15442827  0.49603620   0   0
#>   0.41435169  0.08178443  0.52824286   0   0
#>   0.26646859 -6.68652174  0.29639059   0   0
#>   0.54090737 -0.29440361  0.84578811   0   0
#>   0.29639059 -7.71196255  0.26336046   0   1
#>   0.52824286  0.15442827  0.49603620   0   0
#>   0.58902320  0.19085984  0.64652547   0   0
#>   0.62853896  0.12916371  0.76095713   0   0
#>   2.26291117  0.22833151  4.40891119   0   1
#>   2.09803860  0.05722946  0.92248880   0   0
#>   1.04496864  0.09436123  0.89866319   0   0
#>   1.03320430  0.17173935  0.83404161   0   0
#>   1.03032921  0.29186844  0.95770325   0   0
#>   0.32912843  4.49375471  0.47232498   0   0
#>   0.46791625  0.95845099  0.48326038   0   0
#>   0.45450001  0.87880114  0.49905771   0   0
#>   0.37575551  0.84981726  0.37327972   0   0
#>   0.30657332  0.94273737  0.25961432   0   0
#>   0.39334424  0.79438216  0.42204897   0   0
#>   0.36576261  0.75396883  0.36958372   0   0
#>   0.34003801  0.71720946  0.35924108   0   0
#>   0.26235279  0.76857777  0.21757521   0   0
#>   0.11940171  0.64926128  0.06303475   0   0
#>   0.05045484  0.82411370 -0.69990294   0   0
#>   0.01076897  0.82411370 -0.69990294   0   0
#>   0.03179524  1.97042055 -0.44281156   0   0
#>   0.10530271  0.93846086  0.11610493   0   0
#>   0.15304975  0.76107659  0.21072126   0   0
#>   0.90646138  0.94140023  0.94337763   0   0
#>   0.88122718  0.86151922  0.89934535   0   0
#>   0.87612607  0.91445953  0.78302895   0   0
#>   0.87509495  1.05543480  0.83583271   0   0
#>   0.77877465  0.12916371  0.76095713   0   0
#>   0.82354686  0.17173935  0.83404161   0   0
#>   0.94257364  0.76903091  1.03874419   0   0
#>   0.99699633  0.81574505  2.50396568   0   0
#>   0.95588467  0.52538711  1.01730439   0   0
#>   0.89538381  0.59658682  0.80788058   0   0
#>  -0.44281156  0.93846086  0.11610493   0   0
#>   0.78666414  0.36151435  0.84156214   0   0
#>   0.72627056  0.40717916  0.71121022   0   0
#>   0.68737348  0.53076401  0.60625724   0   0
#>   0.70211009  0.69733362  0.64570879   0   0
#>   0.76855330  0.60339421  0.78394873   0   0
#>   0.55240000  0.40065097  0.60883404   0   0
#>   0.66931410  0.39120375  0.68126992   0   0
#>   0.69225878  0.29087720  0.75557655   0   0
#>   0.65058915  0.13831867  0.74774516   0   0
#>   0.51698551  0.25597201  0.50233534   0   0
#>   1.73965622  0.81574505  2.50396568   0   0
#>   0.98046869  0.91952435  0.97173686   0   0
#>   0.06303475  0.61796267  0.10364144   0   0
#>  -4.06381812  0.32339412 -0.41971384   0   0
#>   0.78302895  0.84124312  0.76310315   0   0
#>   0.83583271  1.63083619  0.88871863   0   0
#>   0.08406289  0.22088270  0.20466764   0   0
#>   0.04398755  0.13217580  0.11104170   0   0
#>   0.35924108  0.75396883  0.36958372   0   0
#>   0.36348514  0.64628303  0.40577819   0   0
#>   0.22747099  0.64639721  0.24834115   0   0
#>   0.21757521  0.94273737  0.25961432   0   0
#>   0.22103805  0.54005239  0.21772232   0   0
#>   0.20133475  0.61796267  0.10364144   0   0
#>   0.25961432  0.93846086  0.11610493   0   0
#>   0.63770022  0.74955761  0.62222952   0   0
#>   0.95770325  0.25067159  0.85975937   0   0
#>   0.85284673  0.25067159  0.85975937   0   0
#>   0.83567778  0.31679790  0.78628751   0   0
#>   0.47232498  6.32424835  0.48326274   0   1
#>   0.94337763  1.16842470  0.93607018   0   0
#>   0.97173686  1.16842470  0.93607018   0   0
#>   3.48744753  0.81574505  2.50396568   0   0
#>   0.51422370  0.74067281  0.53135597   0   0
#>   0.43189808  0.79438216  0.42204897   0   0
#>   0.24285931  0.45077515  0.36261528   0   0
#>   0.30912153  0.31674802  0.36641303   0   0
#>   0.20713161  0.22174685  0.20683796   0   0
#>   0.88871863  1.16842470  0.93607018   0   0
#>   0.46318332  0.61147679  0.48239378   0   0
#>   0.51090072  0.51204929  0.59595341   0   0
#>   0.51063570  0.41236350  0.59584446   0   0
#>   0.49510336  0.41236350  0.59584446   0   0
#>   0.42368776  0.27566570  0.45918220   0   0
#>   0.39846094  0.31674802  0.36641303   0   0
#>   0.40931765  0.75083682  0.41462147   0   0
#>   0.46084541  0.74067281  0.53135597   0   0
#>   0.48239378  0.58845735  0.55908014   0   0
#>   0.74774516  0.25426692  0.76698648   0   0
#>   0.83404161  0.18230301  0.83341354   0   0
#>   0.21772232  0.46145347  0.25784535   0   0
#>   0.50233534  0.27566570  0.45918220   0   0
#>   0.85975937  0.22567131  0.83686496   0   0
#>   0.11291132  0.13217580  0.11104170   0   0
#>   0.78160460  0.40717916  0.71121022   0   0
#>   0.55908014  0.69733362  0.64570879   0   0
#>   0.59595341  0.41192907  0.60053537   0   0
#>   0.68126992  0.40065097  0.60883404   0   0
#>   0.64570879  0.71595441  0.64871561   0   0
#>   0.60053537  0.41236350  0.59584446   0   0
#>   0.36641303  0.19207564  0.26724683   0   0
#>   0.62222952  0.74067281  0.53135597   0   0
#>   0.20466764  0.22174685  0.20683796   0   0
#>   0.89638670  0.76859317  0.92220549   0   0
#>   0.92220549  0.71103589  0.88612349   0   0
#>   0.88612349  0.69992280  0.78542945   0   0
#>   0.78542945  0.71311835  0.77220948   0   0
#>   0.77220948  0.80524036  0.78152181   0   0
#>   0.78152181  0.81477298  0.89638670   0   0
#>   0.87454503  0.52832677  0.91664684   0   0
#>   0.91664684  0.50285770  0.94104424   0   0
#>   0.94104424  0.45905536  0.93743169   0   0
#>   0.93743169  0.42332928  0.92387233   0   0
#>   0.92387233  0.41936270  0.90159728   0   0
#>   0.90159728  0.47145069  0.82016033   0   0
#>   0.82016033  0.47904231  0.82267503   0   0
#>   0.82267503  0.52537424  0.87454503   0   0
#>   0.11949817  0.42181182  0.06824122   0   0
#>   0.06824122  0.44494030  0.08792738   0   0
#>   0.08792738  0.45954316  0.13837876   0   0
#>   0.13837876  0.42303723  0.20671619   0   0
#>   0.20671619  0.40581937  0.20857869   0   0
#>   0.20857869  0.32824707  0.17101471   0   0
#>   0.17101471  0.32323382  0.11949817   0   0
#>   0.59776213  0.94935157  0.60317372   0   0
#>   0.60317372  1.00207033  0.61928234   0   0
#>   0.61928234  0.93108969  0.76133422   0   0
#>   0.76133422  0.86904246  0.61948088   0   0
#>   0.61948088  0.89821837  0.59776213   0   0
#>   0.33136699  0.56256676  0.27791031   0   0
#>   0.27791031  0.60788985  0.30189116   0   0
#>   0.30189116  0.61964440  0.37643177   0   0
#>   0.37643177  0.46134389  0.35203863   0   0
#>   0.35203863  0.46728257  0.33136699   0   0
#>   0.21270042  0.13893835  0.28320955   0   0
#>   0.28320955  0.15191553  0.37908220   0   0
#>   0.37908220  0.07630353  0.41532775   0   0
#>   0.41532775  0.02959279  0.41435169   0   0
#>   0.41435169 -0.31456706  0.26646859   0   0
#>   0.26646859 -0.04724438  0.21270042   0   0
#>   0.54090737 -6.68652174  0.29639059   0   0
#>   0.52824286  0.05086688  0.58902320   0   0
#>   0.58902320 -0.08553273  0.62853896   0   0
#>   0.62853896 -3.08840640  0.54090737   0   0
#>   2.26291117  0.22925336  2.09803860   0   0
#>   2.09803860  0.28006115  1.04496864   0   0
#>   1.04496864  0.28363957  1.03320430   0   0
#>   1.03320430  0.28523001  1.03032921   0   0
#>   0.32912843  0.94077432  0.46791625   0   0
#>   0.46791625  0.91194502  0.45450001   0   0
#>   0.45450001  0.90597645  0.37575551   0   0
#>   0.37575551  0.98154292  0.30657332   0   0
#>   0.30657332  1.22548067  0.32912843   0   0
#>   0.39334424  0.77776196  0.36576261   0   0
#>   0.36576261  0.76287202  0.34003801   0   0
#>   0.34003801  0.78170882  0.26235279   0   0
#>   0.26235279  0.84981726  0.37327972   0   0
#>   0.37327972  0.81617625  0.39334424   0   0
#>   0.11940171  0.78522817  0.05045484   0   0
#>   0.05045484  0.85872891  0.01076897   0   0
#>   0.01076897  0.87528654  0.03179524   0   0
#>   0.03179524  0.89684123  0.10530271   0   0
#>   0.10530271  0.75825469  0.15304975   0   0
#>   0.15304975  0.73465468  0.11940171   0   0
#>   0.90646138  0.87851532  0.88122718   0   0
#>   0.88122718  0.88140201  0.87612607   0   0
#>   0.87612607  0.99473577  0.87509495   0   0
#>   0.87509495  0.95915456  0.90646138   0   0
#>   0.84578811  0.12792901  0.77877465   0   0
#>   0.77877465  0.14066325  0.82354686   0   0
#>   0.82354686  0.09436123  0.89866319   0   0
#>   0.89866319  0.05722946  0.92248880   0   0
#>   0.92248880 -0.29440361  0.84578811   0   0
#>   0.94257364  0.63508920  0.99699633   0   0
#>   0.99699633  0.59876944  0.95588467   0   0
#>   0.95588467  0.58228045  0.92125990   0   0
#>   0.92125990  0.61419959  0.89538381   0   0
#>   0.89538381  0.67954567  0.90337985   0   0
#>   0.90337985  0.66703667  0.94257364   0   0
#>  -0.69990294  0.70336610 -7.77386881   0   1
#>   0.78666414  0.44165833  0.72627056   0   0
#>   0.72627056  0.53488888  0.68737348   0   0
#>   0.68737348  0.55537108  0.70211009   0   0
#>   0.70211009  0.58753490  0.76855330   0   0
#>   0.55240000  0.33343610  0.66931410   0   0
#>   0.66931410  0.26078688  0.69225878   0   0
#>   0.69225878  0.19139589  0.65058915   0   0
#>   0.65058915  0.19085984  0.64652547   0   0
#>   0.64652547  0.25620059  0.51698551   0   0
#>   0.51698551  0.29581626  0.55240000   0   0
#>   1.73965622  0.76903091  1.03874419   0   0
#>   0.89934535  0.90912108  0.98046869   0   0
#>   0.98046869  0.82512692  1.73965622   0   0
#>   0.78394873  0.59658682  0.80788058   0   0
#>   0.06303475  0.62973532  0.01783472   0   0
#>   0.01783472  0.35201411 -4.06381812   0   0
#>  -4.06381812  0.31077153 -5.67894222   0   1
#>   0.78302895  0.92780588  0.77891372   0   0
#>   0.77891372  1.05543480  0.83583271   0   0
#>   0.08406289  0.24060804  0.04398755   0   0
#>   0.04398755  0.32339412 -0.41971384   0   0
#>   0.35924108  0.68461883  0.36348514   0   0
#>   0.36348514  0.73870202  0.22747099   0   0
#>   0.22747099  0.76107659  0.21072126   0   0
#>   0.21072126  0.76857777  0.21757521   0   0
#>   0.22103805  0.59219142  0.20133475   0   0
#>   0.20133475  0.64639721  0.24834115   0   0
#>   0.76310315  0.82522832  0.63770022   0   0
#>   0.95770325  0.32934965  0.85284673   0   0
#>   0.85284673  0.34063601  0.83567778   0   0
#>   0.83567778  0.36151435  0.84156214   0   0
#>   0.47232498  1.33524129  0.58038378   0   0
#>   0.58038378  0.95845099  0.48326038   0   0
#>   0.94337763  0.91952435  0.97173686   0   0
#>   1.01730439  0.83872902  3.48744753   0   0
#>   3.48744753  0.92766549  5.42234026   0   1
#>   0.49905771  0.87399790  0.51453204   0   0
#>   0.51453204  0.76377857  0.51422370   0   0
#>   0.51422370  0.79482010  0.43189808   0   0
#>   0.43189808  0.87880114  0.49905771   0   0
#>   0.24285931  0.32517575  0.30912153   0   0
#>   0.30912153  0.29423880  0.20713161   0   0
#>   0.36259740  0.51072983  0.46318332   0   0
#>   0.46318332  0.47096005  0.51090072   0   0
#>   0.51090072  0.46437589  0.51063570   0   0
#>   0.51063570  0.40341576  0.49510336   0   0
#>   0.49510336  0.31518832  0.42368776   0   0
#>   0.42368776  0.32740618  0.39846094   0   0
#>   0.39846094  0.45077515  0.36261528   0   0
#>   0.36261528  0.45084096  0.36259740   0   0
#>   0.40577819  0.64686253  0.40931765   0   0
#>   0.40931765  0.63739156  0.46084541   0   0
#>   0.46084541  0.61147679  0.48239378   0   0
#>   0.42204897  0.75083682  0.41462147   0   0
#>   0.41462147  0.75396883  0.36958372   0   0
#>   0.74774516  0.12916371  0.76095713   0   0
#>   0.21772232  0.50688327  0.15389382   0   0
#>   0.15389382  0.61796267  0.10364144   0   0
#>   0.49603620  0.25597201  0.50233534   0   0
#>   0.83341354  0.25426692  0.76698648   0   0
#>   0.76698648  0.29087720  0.75557655   0   0
#>   0.75557655  0.31686731  0.78160460   0   0
#>   0.78160460  0.31679790  0.78628751   0   0
#>   0.78628751  0.22567131  0.83686496   0   0
#>   0.83686496  0.18230301  0.83341354   0   0
#>   0.55908014  0.53076401  0.60625724   0   0
#>   0.60625724  0.51204929  0.59595341   0   0
#>   0.68126992  0.40717916  0.71121022   0   0
#>   0.60883404  0.41192907  0.60053537   0   0
#>   0.64871561  0.74955761  0.62222952   0   0
#>   0.20466764  0.13217580  0.11104170   0   0
#>   0.26724683  0.22174685  0.20683796   0   0
#>  -0.44281156  6.37422882 -2.74619335   0   1
#>   0.88871863  5.58949242  1.12491160   0   1
#>   0.11291132 -0.42797463 -8.52874557   0   1
#> 
#> $x
#>              [,1]        [,2]
#>   [1,] 0.77292205 0.841992785
#>   [2,] 0.47613952 0.876843679
#>   [3,] 0.38651946 0.146282588
#>   [4,] 0.92106768 0.686769649
#>   [5,] 0.55474373 0.317934130
#>   [6,] 0.09776455 0.355236889
#>   [7,] 0.03362977 0.504494081
#>   [8,] 0.43163733 0.994097625
#>   [9,] 0.97616690 0.436057792
#>  [10,] 0.78391097 0.342388791
#>  [11,] 0.83700197 0.054918631
#>  [12,] 0.93360865 0.884801355
#>  [13,] 0.03721209 0.828764877
#>  [14,] 0.63086086 0.939537020
#>  [15,] 0.90674226 0.000000000
#>  [16,] 0.50905163 0.777485326
#>  [17,] 0.30249758 0.640521399
#>  [18,] 0.83919312 0.960526048
#>  [19,] 0.53319387 0.825880930
#>  [20,] 0.76542002 0.002412271
#>  [21,] 0.93344216 0.866502335
#>  [22,] 0.33765631 0.052308369
#>  [23,] 0.74309293 0.332491397
#>  [24,] 0.64101545 0.856550656
#>  [25,] 0.69828308 0.961055330
#>  [26,] 0.59125601 0.248926997
#>  [27,] 0.85208795 0.300528554
#>  [28,] 0.49321211 0.020933667
#>  [29,] 0.13408639 0.979741594
#>  [30,] 0.88313867 0.703359952
#>  [31,] 0.57444788 0.869949466
#>  [32,] 0.35753106 0.897964607
#>  [33,] 0.98767239 0.459659903
#>  [34,] 0.92581855 0.909695656
#>  [35,] 0.46598235 1.000000000
#>  [36,] 0.81970856 0.478421328
#>  [37,] 0.34497874 0.232067155
#>  [38,] 0.99673580 0.724579585
#>  [39,] 0.97793459 0.949897176
#>  [40,] 1.00000000 0.178298366
#>  [41,] 0.45231085 0.464974148
#>  [42,] 0.54014426 0.412678123
#>  [43,] 0.75758024 0.379060885
#>  [44,] 0.08148396 0.669676058
#>  [45,] 0.18248466 0.918311515
#>  [46,] 0.41802574 0.839673448
#>  [47,] 0.55510371 0.171495191
#>  [48,] 0.12933585 0.553177557
#>  [49,] 0.27895331 0.869876810
#>  [50,] 0.03137898 0.025183476
#>  [51,] 0.24590523 0.789281125
#>  [52,] 0.84708485 0.835838023
#>  [53,] 0.23812439 0.887050148
#>  [54,] 0.51949539 0.520968688
#>  [55,] 0.63917208 0.193672214
#>  [56,] 0.32464736 0.710653799
#>  [57,] 0.61211685 0.634236405
#>  [58,] 0.45266903 0.642344555
#>  [59,] 0.39722039 0.275370898
#>  [60,] 0.74908039 0.378469791
#>  [61,] 0.38757072 0.791380765
#>  [62,] 0.09523562 0.476262652
#>  [63,] 0.43193173 0.544956534
#>  [64,] 0.23471943 0.753390002
#>  [65,] 0.52433176 0.263726135
#>  [66,] 0.66339389 0.300800737
#>  [67,] 0.56585112 0.970495930
#>  [68,] 0.80989626 0.004717148
#>  [69,] 0.84567710 0.445948479
#>  [70,] 0.78567784 0.715806362
#>  [71,] 0.93919269 0.515509270
#>  [72,] 0.64322461 0.712534895
#>  [73,] 0.88073431 0.181852425
#>  [74,] 0.22170483 0.139540015
#>  [75,] 0.26590234 0.256053671
#>  [76,] 0.81950737 0.550338442
#>  [77,] 0.68732142 0.972680779
#>  [78,] 0.40028742 0.273560897
#>  [79,] 0.19777873 0.341699119
#>  [80,] 0.75443116 0.359452319
#>  [81,] 0.14667600 0.018212048
#>  [82,] 0.49357257 0.203470477
#>  [83,] 0.52549889 0.106055785
#>  [84,] 0.66700789 0.565248465
#>  [85,] 0.15580548 0.201977054
#>  [86,] 0.98171725 0.976294476
#>  [87,] 0.44867997 0.555179850
#>  [88,] 0.02619324 0.759321770
#>  [89,] 0.38413270 0.549202004
#>  [90,] 0.26629962 0.157981594
#>  [91,] 0.34177814 0.627881960
#>  [92,] 0.97743970 0.934522087
#>  [93,] 0.00000000 0.999363105
#>  [94,] 0.23137553 0.773540337
#>  [95,] 0.13659043 0.436231728
#>  [96,] 0.91999591 0.030582210
#>  [97,] 0.57858968 0.985715566
#>  [98,] 0.97262248 0.518046031
#>  [99,] 0.74541082 0.450406774
#> [100,] 0.27675581 0.844865352
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.772922,0.8419928) [6]: 18 24 25 52 70 72 
#> 2: (0.4761395,0.8768437) [8]: 8 16 19 31 32 35 46 67 
#> 3: (0.3865195,0.1462826) [7]: 22 28 37 78 82 83 90 
#> 4: (0.9210677,0.6867696) [5]: 30 38 71 76 98 
#> 5: (0.5547437,0.3179341) [5]: 26 42 65 66 78 
#> 6: (0.09776455,0.3552369) [6]: 7 50 62 79 85 95 
#> 7: (0.03362977,0.5044941) [7]: 6 44 48 50 62 88 93 
#> 8: (0.4316373,0.9940976) [7]: 2 29 32 35 45 53 93 
#> 9: (0.9761669,0.4360578) [5]: 27 33 40 69 71 
#> 10: (0.783911,0.3423888) [5]: 23 27 43 69 80 
#> 11: (0.837002,0.05491863) [6]: 15 20 55 68 73 96 
#> 12: (0.9336087,0.8848014) [4]: 21 34 52 92 
#> 13: (0.03721209,0.8287649) [5]: 29 45 88 93 94 
#> 14: (0.6308609,0.939537) [6]: 24 25 31 67 77 97 
#> 15: (0.9067423,0) [5]: 11 20 40 68 96 
#> 16: (0.5090516,0.7774853) [7]: 2 19 46 57 58 61 72 
#> 17: (0.3024976,0.6405214) [6]: 44 48 56 64 89 91 
#> 18: (0.8391931,0.960526) [6]: 1 25 34 52 77 86 
#> 19: (0.5331939,0.8258809) [5]: 2 16 24 31 72 
#> 20: (0.76542,0.002412271) [7]: 11 15 28 55 68 81 83 
#> 21: (0.9334422,0.8665023) [5]: 12 30 38 52 92 
#> 22: (0.3376563,0.05230837) [5]: 3 28 74 81 90 
#> 23: (0.7430929,0.3324914) [7]: 10 27 55 60 66 73 80 
#> 24: (0.6410155,0.8565507) [6]: 1 14 19 25 31 72 
#> 25: (0.6982831,0.9610553) [5]: 1 14 18 24 77 
#> 26: (0.591256,0.248927) [5]: 5 47 55 65 66 
#> 27: (0.8520879,0.3005286) [6]: 9 10 23 40 69 73 
#> 28: (0.4932121,0.02093367) [5]: 3 20 22 81 83 
#> 29: (0.1340864,0.9797416) [4]: 8 13 45 93 
#> 30: (0.8831387,0.70336) [6]: 4 21 38 52 70 76 
#> 31: (0.5744479,0.8699495) [5]: 2 14 19 24 67 
#> 32: (0.3575311,0.8979646) [7]: 2 8 46 49 53 61 100 
#> 33: (0.9876724,0.4596599) [5]: 9 38 40 71 98 
#> 34: (0.9258186,0.9096957) [6]: 12 18 39 52 86 92 
#> 35: (0.4659823,1) [6]: 2 8 67 86 93 97 
#> 36: (0.8197086,0.4784213) [4]: 69 71 76 99 
#> 37: (0.3449787,0.2320672) [5]: 3 59 75 78 90 
#> 38: (0.9967358,0.7245796) [8]: 4 21 30 33 40 86 92 98 
#> 39: (0.9779346,0.9498972) [3]: 34 86 92 
#> 40: (1,0.1782984) [7]: 9 15 27 33 38 73 96 
#> 41: (0.4523108,0.4649741) [8]: 42 54 59 63 78 79 87 89 
#> 42: (0.5401443,0.4126781) [8]: 5 41 54 60 66 78 84 99 
#> 43: (0.7575802,0.3790609) [5]: 10 60 69 80 99 
#> 44: (0.08148396,0.6696761) [6]: 7 17 48 64 88 94 
#> 45: (0.1824847,0.9183115) [5]: 8 13 29 53 94 
#> 46: (0.4180257,0.8396734) [4]: 2 16 32 61 
#> 47: (0.5551037,0.1714952) [5]: 26 55 65 82 83 
#> 48: (0.1293358,0.5531776) [6]: 7 17 44 62 89 95 
#> 49: (0.2789533,0.8698768) [3]: 32 53 100 
#> 50: (0.03137898,0.02518348) [5]: 6 7 81 85 93 
#> 51: (0.2459052,0.7892811) [6]: 53 56 61 64 94 100 
#> 52: (0.8470848,0.835838) [7]: 1 12 18 21 30 34 70 
#> 53: (0.2381244,0.8870501) [7]: 8 32 45 49 51 94 100 
#> 54: (0.5194954,0.5209687) [6]: 41 42 57 58 84 87 
#> 55: (0.6391721,0.1936722) [8]: 11 20 23 26 47 66 73 83 
#> 56: (0.3246474,0.7106538) [6]: 17 51 58 61 64 91 
#> 57: (0.6121168,0.6342364) [5]: 16 54 58 72 84 
#> 58: (0.452669,0.6423446) [8]: 16 54 56 57 61 87 89 91 
#> 59: (0.3972204,0.2753709) [5]: 37 41 75 78 79 
#> 60: (0.7490804,0.3784698) [6]: 23 42 43 66 80 99 
#> 61: (0.3875707,0.7913808) [7]: 16 32 46 51 56 58 100 
#> 62: (0.09523562,0.4762627) [4]: 6 7 48 95 
#> 63: (0.4319317,0.5449565) [3]: 41 87 89 
#> 64: (0.2347194,0.75339) [5]: 17 44 51 56 94 
#> 65: (0.5243318,0.2637261) [5]: 5 26 47 78 82 
#> 66: (0.6633939,0.3008007) [6]: 5 23 26 42 55 60 
#> 67: (0.5658511,0.9704959) [5]: 2 14 31 35 97 
#> 68: (0.8098963,0.004717148) [3]: 11 15 20 
#> 69: (0.8456771,0.4459485) [7]: 9 10 27 36 43 71 99 
#> 70: (0.7856778,0.7158064) [6]: 1 30 52 72 76 84 
#> 71: (0.9391927,0.5155093) [7]: 4 9 33 36 69 76 98 
#> 72: (0.6432246,0.7125349) [7]: 1 16 19 24 57 70 84 
#> 73: (0.8807343,0.1818524) [6]: 11 23 27 40 55 96 
#> 74: (0.2217048,0.13954) [4]: 22 81 85 90 
#> 75: (0.2659023,0.2560537) [5]: 37 59 79 85 90 
#> 76: (0.8195074,0.5503384) [7]: 4 30 36 70 71 84 99 
#> 77: (0.6873214,0.9726808) [5]: 14 18 25 86 97 
#> 78: (0.4002874,0.2735609) [8]: 3 5 37 41 42 59 65 82 
#> 79: (0.1977787,0.3416991) [7]: 6 41 59 75 85 89 95 
#> 80: (0.7544312,0.3594523) [4]: 10 23 43 60 
#> 81: (0.146676,0.01821205) [6]: 20 22 28 50 74 85 
#> 82: (0.4935726,0.2034705) [5]: 3 47 65 78 83 
#> 83: (0.5254989,0.1060558) [6]: 3 20 28 47 55 82 
#> 84: (0.6670079,0.5652485) [7]: 42 54 57 70 72 76 99 
#> 85: (0.1558055,0.2019771) [7]: 6 50 74 75 79 81 90 
#> 86: (0.9817172,0.9762945) [8]: 18 34 35 38 39 77 92 97 
#> 87: (0.44868,0.5551798) [5]: 41 54 58 63 89 
#> 88: (0.02619324,0.7593218) [5]: 7 13 44 93 94 
#> 89: (0.3841327,0.549202) [9]: 17 41 48 58 63 79 87 91 95 
#> 90: (0.2662996,0.1579816) [6]: 3 22 37 74 75 85 
#> 91: (0.3417781,0.627882) [4]: 17 56 58 89 
#> 92: (0.9774397,0.9345221) [6]: 12 21 34 38 39 86 
#> 93: (0,0.9993631) [7]: 7 8 13 29 35 50 88 
#> 94: (0.2313755,0.7735403) [7]: 13 44 45 51 53 64 88 
#> 95: (0.1365904,0.4362317) [5]: 6 48 62 79 89 
#> 96: (0.9199959,0.03058221) [4]: 11 15 40 73 
#> 97: (0.5785897,0.9857156) [5]: 14 35 67 77 86 
#> 98: (0.9726225,0.518046) [4]: 4 33 38 71 
#> 99: (0.7454108,0.4504068) [7]: 36 42 43 60 69 76 84 
#> 100: (0.2767558,0.8448654) [5]: 32 49 51 53 61 
#> number of nodes: 100 
#> number of arcs: 288 
#> number of boundary nodes: 9 
#> boundary nodes:  15 20 35 38 40 50 81 86 93 
#> number of triangles: 189 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.124937516 0.140911031 0.119167144 0.144032704 0.142490796 0.134827371 
#>                                                                         
#> 0.060364967 0.100913309 0.104135780 0.034848507 0.121380790 0.084008728 
#>                                                                         
#> 0.110240983 0.054083103 0.127509249 0.158688303 0.091039613 0.102512987 
#>                                                                         
#> 0.116682415 0.069143189 0.108069767 0.124650097 0.033525906 0.207936435 
#>                                                                         
#> 0.115562026 0.165724243 0.124433606 0.068541005 0.088852456 0.166454261 
#>                                                                         
#> 0.197265151 0.216189555 0.145890399 0.112607379 0.057556020 0.067766468 
#>                                                                         
#> 0.479315889 0.241466205 0.974684867 0.084135180 0.125943258 0.105325293 
#>                                                                         
#> 0.465982784 0.135514438 0.078205163 0.063820470 0.119904454 0.281631469 
#>                                                                         
#> 0.073955618 0.116549858 0.145561168 0.191880678 0.110611960 0.019859824 
#>                                                                         
#> 0.029248021 0.113584928 0.229170041 0.044535929 0.096960809 0.033330630 
#>                                                                         
#> 0.156282281 0.241851230 0.057280841 0.107953443 0.091639979 0.081009073 
#>                                                                         
#> 0.205674354 0.152803831 0.015978391 0.109510260 0.019847126 0.067902650 
#>                                                                         
#> 0.167990229 0.057093711 0.162633276 0.159653836 0.084251684 0.157968629 
#>                                                                         
#> 0.089355717 0.084525978 0.099566144 0.174611378 0.254827867 0.294418005 
#>                                                                         
#> 0.086948647 0.143520843 0.346546798 0.137296457 0.210827402 0.048257509 
#>                                                                         
#> 0.142652727 0.019755885 0.115648859 0.109832676 0.122084551 0.097228905 
#>                                                                         
#> 0.086944295 0.119318636 0.168890703 0.044293618 0.025107809 0.123046922 
#>                                                                         
#> 0.546290971 0.015383052 0.026666945 0.403237638 0.124405131 0.100365344 
#>                                                                         
#> 0.003561284 0.132731256 0.098072881 0.041990832 0.110241655 0.078646210 
#>                                                                         
#> 0.019621922 0.047987200 0.278900076 0.210181854 0.072030516 0.139052627 
#>                                                                         
#> 0.154015058 0.020425907 0.113710264 0.067652587 0.272101930 0.057200906 
#>                                                                         
#> 0.183991644 0.102353269 0.088161005 0.087255936 0.111830034 0.149194293 
#>                                                                         
#> 0.064823492 0.109434773 0.153226626 0.118930884 0.135801287 0.140523537 
#>                                                                         
#> 0.132707502 0.183250567 0.126829506 0.074417757 0.098549800 0.129687756 
#>                                                                         
#> 0.123574461 0.125415042 0.120474317 0.068984326 0.104667518 0.076500982 
#>                                                                         
#> 0.105918651 0.164607636 0.144684064 0.121370600 0.128020798 0.095313291 
#>                                                                         
#> 0.120787736 0.172216822 0.176424367 0.084588747 0.041398647 0.170082247 
#>                                                                         
#> 0.062156210 0.078071320 0.109992777 0.095862238 0.160703877 0.163882051 
#>                                                                         
#> 0.100926246 0.089819892 0.121052182 0.162453006 0.336663480 0.496010396 
#>                                                                         
#> 0.107376603 0.171974152 0.254936175 0.431669444 0.297897060 0.260423859 
#>                                                                         
#> 0.221147960 0.026257111 0.087633498 0.130864110 0.183749183 0.258858915 
#>                                                                         
#> 0.045145869 0.034062045 0.042000837 0.080002373 0.120580528 0.088774398 
#>                                                                         
#> 0.057051802 0.088768036 0.086488473 0.134256119 0.241638619 0.026084705 
#>                                                                         
#> 0.099417182 0.018299778 0.066282065 0.070311879 0.201864282 0.170653798 
#>                                                                         
#> 0.179383939 0.174609551 0.065469871 0.069747645 0.072005007 0.089581539 
#>                                                                         
#> 0.083605337 0.070772831 0.141342829 0.122273036 0.146430957 0.176472923 
#>                                                                         
#> 0.149066920 0.041264009 0.073547036 0.131655608 0.222928270 0.193943111 
#>                                                                         
#> 0.122488880 0.152357306 0.100437517 0.143393765 0.112098733 0.261350578 
#>                                                                         
#> 0.272837290 0.618945710 0.170721641 0.155396755 0.145100335 0.194000102 
#>                                                                         
#> 0.046366609 0.085768497 0.173407929 0.204052070 0.085455680 0.073137136 
#>                                                                         
#> 0.098252355 0.083446907 0.096665257 0.110736158 0.265074675 0.060294608 
#>                                                                         
#> 0.065819785 0.113509721 0.516279421 0.125107840 0.071917395 0.079403838 
#>                                                                         
#> 0.041579450 0.067855763 0.082634313 0.102223202 0.087459425 0.090278745 
#>                                                                         
#> 0.082537812 0.108363238 0.282813250 0.197444551 0.198356926 0.211718014 
#>                                                                         
#> 0.208705083 0.198424017 0.072376313 0.008520378 0.182360280 0.069343356 
#>                                                                         
#> 0.071824518 0.117170627 0.021421593 0.037593816 0.111277041 0.141681047 
#>                                                                         
#> 0.063571727 0.098078149 0.146315790 0.138556352 0.145105827 0.115640666 
#>                                                                         
#> 0.191703514 0.090780549 0.122660505 0.201214596 0.252162534 0.115507594 
#> 
#> $alpha
#> [1] 0.2830738
#> 
#> attr(,"class")
#> [1] "scree"