Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    59   81 0.709340429 0.01973176 0.619700138 0.07578372  0.657059282
#>     1   81 0.615721078 0.00000000 0.619700138 0.07578372  0.509385667
#>    81    4 0.619700138 0.07578372 0.707173183 0.13061507  0.647667292
#>     4   72 0.707173183 0.13061507 0.736870487 0.14821497  0.673809006
#>    22   72 0.797160985 0.05319628 0.736870487 0.14821497  0.750989267
#>    59   22 0.709340429 0.01973176 0.797160985 0.05319628  0.738276689
#>    50   43 0.589650410 0.14177634 0.554513527 0.25363377  0.469663751
#>     4    3 0.707173183 0.13061507 0.682229523 0.31652569  0.656219522
#>    81   50 0.619700138 0.07578372 0.589650410 0.14177634  0.505608273
#>    20   50 0.394907875 0.03351826 0.589650410 0.14177634  0.486419638
#>     3   43 0.682229523 0.31652569 0.554513527 0.25363377  0.649432910
#>    25    3 0.824223349 0.31075329 0.682229523 0.31652569  0.750462591
#>    11   50 0.381977327 0.05758785 0.589650410 0.14177634  0.461213753
#>    43   64 0.554513527 0.25363377 0.621886674 0.33716991  0.567985210
#>     3   64 0.682229523 0.31652569 0.621886674 0.33716991  0.629816973
#>    25   47 0.824223349 0.31075329 0.792817768 0.33915347  0.751105304
#>    22   58 0.797160985 0.05319628 0.898194888 0.29014199  0.866324497
#>    84   16 0.798933982 0.52755178 0.727182782 0.55113040  0.728942895
#>    74   94 0.503165228 0.28317184 0.520911802 0.34828278  0.437662966
#>    74   43 0.503165228 0.28317184 0.554513527 0.25363377  0.550095782
#>    64   92 0.621886674 0.33716991 0.525967976 0.50499862  0.580432132
#>     1   20 0.615721078 0.00000000 0.394907875 0.03351826  0.418260176
#>    20   11 0.394907875 0.03351826 0.381977327 0.05758785  0.352012969
#>    11   74 0.381977327 0.05758785 0.503165228 0.28317184  0.443969236
#>    58   25 0.898194888 0.29014199 0.824223349 0.31075329  0.833505508
#>    45   96 0.305616941 0.06305787 0.308521147 0.02369735  0.342771982
#>    49   45 0.307486769 0.07547030 0.305616941 0.06305787  0.344033167
#>    61   45 0.289782762 0.03961695 0.305616941 0.06305787  0.264448793
#>    49   56 0.307486769 0.07547030 0.329129774 0.22626872  0.320423176
#>    49   11 0.307486769 0.07547030 0.381977327 0.05758785  0.363424628
#>    58   44 0.898194888 0.29014199 1.000000000 0.27641825  0.952276083
#>    56   74 0.329129774 0.22626872 0.503165228 0.28317184  0.393335404
#>     3   47 0.682229523 0.31652569 0.792817768 0.33915347  0.716291245
#>    96   61 0.308521147 0.02369735 0.289782762 0.03961695  0.245159894
#>    49   95 0.307486769 0.07547030 0.251482332 0.18302935  0.270910137
#>    56    9 0.329129774 0.22626872 0.431204848 0.43835760  0.366723682
#>    97   49 0.228673995 0.05076923 0.307486769 0.07547030  0.261246687
#>    86   49 0.217186468 0.06800602 0.307486769 0.07547030  0.260563447
#>    16   92 0.727182782 0.55113040 0.525967976 0.50499862  0.642161493
#>    58   53 0.898194888 0.29014199 0.997785932 0.34052160  0.914121243
#>    82    9 0.469192487 0.41401921 0.431204848 0.43835760  0.387723528
#>    92    9 0.525967976 0.50499862 0.431204848 0.43835760  0.478982608
#>    92   82 0.525967976 0.50499862 0.469192487 0.41401921  0.551711264
#>    13   60 0.893880853 0.47357700 0.827870384 0.51015465  0.857939495
#>    77   58 0.844983332 0.45061350 0.898194888 0.29014199  0.906991837
#>    77   84 0.844983332 0.45061350 0.798933982 0.52755178  0.732294801
#>    60   84 0.827870384 0.51015465 0.798933982 0.52755178  0.780632605
#>    39   42 0.859283593 0.63357510 0.815455127 0.61663240  0.859518294
#>    63   95 0.148525646 0.15751418 0.251482332 0.18302935  0.185834063
#>    95   56 0.251482332 0.18302935 0.329129774 0.22626872  0.255700652
#>    38   46 0.234362040 0.36816479 0.248661331 0.35036477  0.282456558
#>     8   86 0.132421946 0.08615563 0.217186468 0.06800602  0.175938337
#>    63    8 0.148525646 0.15751418 0.132421946 0.08615563  0.167362716
#>    63   76 0.148525646 0.15751418 0.212248233 0.10665805  0.202576842
#>     8   97 0.132421946 0.08615563 0.228673995 0.05076923  0.162374485
#>    38   98 0.234362040 0.36816479 0.110804258 0.20188864  0.160383537
#>    62   89 0.546057422 0.58358097 0.350499101 0.56361462  0.448999409
#>    46    9 0.248661331 0.35036477 0.431204848 0.43835760  0.362928035
#>    28   48 0.971809774 0.49540176 0.897040431 0.62311924  0.915077125
#>    48   60 0.897040431 0.62311924 0.827870384 0.51015465  0.892211564
#>    13   53 0.893880853 0.47357700 0.997785932 0.34052160  0.952971940
#>    40   26 0.289521198 0.50225981 0.300846034 0.44265988  0.367109850
#>    70   62 0.654023440 0.61074995 0.546057422 0.58358097  0.618663374
#>    95   46 0.251482332 0.18302935 0.248661331 0.35036477  0.190962054
#>    16   70 0.727182782 0.55113040 0.654023440 0.61074995  0.631172570
#>     1   96 0.615721078 0.00000000 0.308521147 0.02369735  0.265442390
#>    96   97 0.308521147 0.02369735 0.228673995 0.05076923  0.198378508
#>     8   98 0.132421946 0.08615563 0.110804258 0.20188864  0.014175746
#>    32   38 0.156832115 0.51313304 0.234362040 0.36816479  0.215934437
#>    40   32 0.289521198 0.50225981 0.156832115 0.51313304  0.219104579
#>    26    9 0.300846034 0.44265988 0.431204848 0.43835760  0.367525364
#>    17   70 0.693789517 0.63765211 0.654023440 0.61074995  0.697471290
#>    99   42 0.808721339 0.61460172 0.815455127 0.61663240  0.780817592
#>    99   84 0.808721339 0.61460172 0.798933982 0.52755178  0.826281381
#>    17   16 0.693789517 0.63765211 0.727182782 0.55113040  0.747777144
#>    89    9 0.350499101 0.56361462 0.431204848 0.43835760  0.437045238
#>    62   93 0.546057422 0.58358097 0.606648308 0.66080265  0.580380824
#>    17   99 0.693789517 0.63765211 0.808721339 0.61460172  0.765407852
#>    48   39 0.897040431 0.62311924 0.859283593 0.63357510  0.861273279
#>    41   68 0.552962856 0.66005668 0.440070369 0.73249351  0.454260705
#>    91   31 0.673160356 0.70695637 0.544674827 0.71015276  0.609397437
#>    93   17 0.606648308 0.66080265 0.693789517 0.63765211  0.654020450
#>    78   31 0.508496468 0.81824680 0.544674827 0.71015276  0.601036688
#>    41   31 0.552962856 0.66005668 0.544674827 0.71015276  0.482286434
#>    32   83 0.156832115 0.51313304 0.011689065 0.30191898  0.105881491
#>    83   98 0.011689065 0.30191898 0.110804258 0.20188864  0.128063133
#>    40   89 0.289521198 0.50225981 0.350499101 0.56361462  0.258243669
#>    89   68 0.350499101 0.56361462 0.440070369 0.73249351  0.376946796
#>    83    6 0.011689065 0.30191898 0.077456580 0.01715316 -0.055238480
#>    33   24 0.635000259 0.89187060 0.511473064 0.82640792  0.609120597
#>    65   23 0.144998001 0.59087816 0.076129187 0.56133640  0.122741739
#>    65   40 0.144998001 0.59087816 0.289521198 0.50225981  0.244521891
#>    27   89 0.244312197 0.69055346 0.350499101 0.56361462  0.329085700
#>    19   68 0.334520205 0.74590880 0.440070369 0.73249351  0.392949611
#>    37   83 0.009460254 0.51706178 0.011689065 0.30191898  0.080350110
#>     5   36 0.067131193 0.64958482 0.074923124 0.61521099  0.124350479
#>    27   52 0.244312197 0.69055346 0.206217915 0.68761442  0.231856491
#>    27   80 0.244312197 0.69055346 0.287373143 0.77539185  0.226705679
#>    68   31 0.440070369 0.73249351 0.544674827 0.71015276  0.499601010
#>    79   68 0.336872915 0.82570746 0.440070369 0.73249351  0.413389547
#>    91   33 0.673160356 0.70695637 0.635000259 0.89187060  0.610956538
#>    23   37 0.076129187 0.56133640 0.009460254 0.51706178  0.082206601
#>    36   37 0.074923124 0.61521099 0.009460254 0.51706178  0.011161332
#>    39    7 0.859283593 0.63357510 0.715310110 0.84107060  0.794679711
#>    52    5 0.206217915 0.68761442 0.067131193 0.64958482  0.138233231
#>    27   19 0.244312197 0.69055346 0.334520205 0.74590880  0.286941771
#>    79   15 0.336872915 0.82570746 0.304253262 0.82568652  0.320559921
#>    80   19 0.287373143 0.77539185 0.334520205 0.74590880  0.324442251
#>    24   66 0.511473064 0.82640792 0.450348281 0.87633355  0.469306728
#>     5   21 0.067131193 0.64958482 0.000000000 0.67951561  0.013406778
#>    73    5 0.159037593 0.82903802 0.067131193 0.64958482  0.118033281
#>    15   80 0.304253262 0.82568652 0.287373143 0.77539185  0.312208926
#>     7   33 0.715310110 0.84107060 0.635000259 0.89187060  0.629533990
#>    85   80 0.265040030 0.80498824 0.287373143 0.77539185  0.291735284
#>    79   66 0.336872915 0.82570746 0.450348281 0.87633355  0.394604760
#>    68   78 0.440070369 0.73249351 0.508496468 0.81824680  0.437165204
#>     2   12 0.332444185 0.85018207 0.315958737 0.87638469  0.306176432
#>    88   69 0.990105321 0.85227660 0.755396654 0.92873593  0.853349011
#>    28   53 0.971809774 0.49540176 0.997785932 0.34052160  2.152375663
#>    53   44 0.997785932 0.34052160 1.000000000 0.27641825 15.745113063
#>    33   69 0.635000259 0.89187060 0.755396654 0.92873593  0.686651358
#>    10   29 0.988221339 0.87088491 0.975270847 0.90741244  0.869944825
#>    12   71 0.315958737 0.87638469 0.266613992 0.86503245  0.295010487
#>    71   85 0.266613992 0.86503245 0.265040030 0.80498824  0.274380842
#>     2   66 0.332444185 0.85018207 0.450348281 0.87633355  0.384752677
#>    14   30 0.549674654 1.00000000 0.543791747 0.99239376  0.601930706
#>    21   37 0.000000000 0.67951561 0.009460254 0.51706178 -0.005151834
#>    71   73 0.266613992 0.86503245 0.159037593 0.82903802  0.216415872
#>    29   69 0.975270847 0.90741244 0.755396654 0.92873593  0.858382738
#>    29   54 0.975270847 0.90741244 0.981361986 0.93239637  0.876630550
#>    85   73 0.265040030 0.80498824 0.159037593 0.82903802  0.200042027
#>    67   66 0.318970681 0.90934529 0.450348281 0.87633355  0.412252645
#>    57   55 0.751155912 0.96380903 0.781936001 0.99109489  0.669316775
#>    57   69 0.751155912 0.96380903 0.755396654 0.92873593  0.790225457
#>    57   14 0.751155912 0.96380903 0.549674654 1.00000000  0.652478959
#>    90   73 0.048759188 0.83308685 0.159037593 0.82903802  0.105317848
#>    75  100 0.244334309 0.97161912 0.151144603 0.91913196  0.217451156
#>    67   30 0.318970681 0.90934529 0.543791747 0.99239376  0.388093552
#>    67   75 0.318970681 0.90934529 0.244334309 0.97161912  0.264790891
#>    90  100 0.048759188 0.83308685 0.151144603 0.91913196  0.093996786
#>    18   87 0.135601349 0.96263417 0.135717350 0.93626803  0.167748183
#>    18   35 0.135601349 0.96263417 0.108310791 0.97121336  0.116436689
#>    75   18 0.244334309 0.97161912 0.135601349 0.96263417  0.190737983
#>    35   90 0.108310791 0.97121336 0.048759188 0.83308685  0.064143708
#>    14   51 0.549674654 1.00000000 0.022316312 0.98955543  0.256162444
#>    21   90 0.000000000 0.67951561 0.048759188 0.83308685 -0.290387462
#>    51   35 0.022316312 0.98955543 0.108310791 0.97121336  0.174765089
#>     4   59 0.707173183 0.13061507 0.709340429 0.01973176  0.681334234
#>    20   81 0.394907875 0.03351826 0.619700138 0.07578372  0.509385667
#>    50    4 0.589650410 0.14177634 0.707173183 0.13061507  0.647667292
#>     3   72 0.682229523 0.31652569 0.736870487 0.14821497  0.673809006
#>     4   22 0.707173183 0.13061507 0.797160985 0.05319628  0.738276689
#>    50    3 0.589650410 0.14177634 0.682229523 0.31652569  0.656219522
#>    74   50 0.503165228 0.28317184 0.589650410 0.14177634  0.461213753
#>    94   64 0.520911802 0.34828278 0.621886674 0.33716991  0.567985210
#>    72   58 0.736870487 0.14821497 0.898194888 0.29014199  0.866324497
#>    47   16 0.792817768 0.33915347 0.727182782 0.55113040  0.728942895
#>    82   94 0.469192487 0.41401921 0.520911802 0.34828278  0.437662966
#>    94   43 0.520911802 0.34828278 0.554513527 0.25363377  0.550095782
#>    94   92 0.520911802 0.34828278 0.525967976 0.50499862  0.580432132
#>    96   20 0.308521147 0.02369735 0.394907875 0.03351826  0.418260176
#>    96   11 0.308521147 0.02369735 0.381977327 0.05758785  0.352012969
#>    77   25 0.844983332 0.45061350 0.824223349 0.31075329  0.856039989
#>    72   25 0.736870487 0.14821497 0.824223349 0.31075329  0.833505508
#>    11   45 0.381977327 0.05758785 0.305616941 0.06305787  0.344033167
#>    97   45 0.228673995 0.05076923 0.305616941 0.06305787  0.264448793
#>    56   11 0.329129774 0.22626872 0.381977327 0.05758785  0.363424628
#>    44   22 1.000000000 0.27641825 0.797160985 0.05319628  0.929356885
#>    82   74 0.469192487 0.41401921 0.503165228 0.28317184  0.393335404
#>    97   61 0.228673995 0.05076923 0.289782762 0.03961695  0.245159894
#>    76   95 0.212248233 0.10665805 0.251482332 0.18302935  0.270910137
#>    76   49 0.212248233 0.10665805 0.307486769 0.07547030  0.260563447
#>    77   60 0.844983332 0.45061350 0.827870384 0.51015465  0.857939495
#>    13   58 0.893880853 0.47357700 0.898194888 0.29014199  0.906991837
#>    47   84 0.792817768 0.33915347 0.798933982 0.52755178  0.732294801
#>    60   42 0.827870384 0.51015465 0.815455127 0.61663240  0.859518294
#>    98   95 0.110804258 0.20188864 0.251482332 0.18302935  0.185834063
#>    46   56 0.248661331 0.35036477 0.329129774 0.22626872  0.255700652
#>    26   46 0.300846034 0.44265988 0.248661331 0.35036477  0.282456558
#>    16    3 0.727182782 0.55113040 0.682229523 0.31652569  0.689645006
#>    76   86 0.212248233 0.10665805 0.217186468 0.06800602  0.175938337
#>    76    8 0.212248233 0.10665805 0.132421946 0.08615563  0.167362716
#>    86   97 0.217186468 0.06800602 0.228673995 0.05076923  0.162374485
#>    46   98 0.248661331 0.35036477 0.110804258 0.20188864  0.160383537
#>    92   89 0.525967976 0.50499862 0.350499101 0.56361462  0.448999409
#>    13   48 0.893880853 0.47357700 0.897040431 0.62311924  0.915077125
#>    89   26 0.350499101 0.56361462 0.300846034 0.44265988  0.367109850
#>    92   62 0.525967976 0.50499862 0.546057422 0.58358097  0.618663374
#>    98   63 0.110804258 0.20188864 0.148525646 0.15751418  0.078140335
#>    92   70 0.525967976 0.50499862 0.654023440 0.61074995  0.631172570
#>     6   96 0.077456580 0.01715316 0.308521147 0.02369735  0.265442390
#>     6   97 0.077456580 0.01715316 0.228673995 0.05076923  0.198378508
#>     6    8 0.077456580 0.01715316 0.132421946 0.08615563  0.158641662
#>     6   98 0.077456580 0.01715316 0.110804258 0.20188864  0.014175746
#>    26   38 0.300846034 0.44265988 0.234362040 0.36816479  0.215934437
#>    26   32 0.300846034 0.44265988 0.156832115 0.51313304  0.219104579
#>    91   42 0.673160356 0.70695637 0.815455127 0.61663240  0.780817592
#>    42   84 0.815455127 0.61663240 0.798933982 0.52755178  0.826281381
#>    93   70 0.606648308 0.66080265 0.654023440 0.61074995  0.598021062
#>    99   16 0.808721339 0.61460172 0.727182782 0.55113040  0.747777144
#>    41   93 0.552962856 0.66005668 0.606648308 0.66080265  0.580380824
#>    91   39 0.673160356 0.70695637 0.859283593 0.63357510  0.793331061
#>    91   99 0.673160356 0.70695637 0.808721339 0.61460172  0.765407852
#>    60   39 0.827870384 0.51015465 0.859283593 0.63357510  0.861273279
#>    62   68 0.546057422 0.58358097 0.440070369 0.73249351  0.454260705
#>    93   31 0.606648308 0.66080265 0.544674827 0.71015276  0.609397437
#>    91   17 0.673160356 0.70695637 0.693789517 0.63765211  0.654020450
#>    24   31 0.511473064 0.82640792 0.544674827 0.71015276  0.601036688
#>    38   83 0.234362040 0.36816479 0.011689065 0.30191898  0.105881491
#>    32   23 0.156832115 0.51313304 0.076129187 0.56133640  0.122741739
#>    27   40 0.244312197 0.69055346 0.289521198 0.50225981  0.244521891
#>    19   89 0.334520205 0.74590880 0.350499101 0.56361462  0.329085700
#>    34   68 0.332966393 0.81894979 0.440070369 0.73249351  0.392949611
#>    65   36 0.144998001 0.59087816 0.074923124 0.61521099  0.124350479
#>    65   52 0.144998001 0.59087816 0.206217915 0.68761442  0.231856491
#>    88   48 0.990105321 0.85227660 0.897040431 0.62311924  0.911995455
#>    66   68 0.450348281 0.87633355 0.440070369 0.73249351  0.413389547
#>    34   79 0.332966393 0.81894979 0.336872915 0.82570746  0.395752986
#>    31   33 0.544674827 0.71015276 0.635000259 0.89187060  0.610956538
#>    32   37 0.156832115 0.51313304 0.009460254 0.51706178  0.082206601
#>    91    7 0.673160356 0.70695637 0.715310110 0.84107060  0.794679711
#>    65    5 0.144998001 0.59087816 0.067131193 0.64958482  0.138233231
#>    34   15 0.332966393 0.81894979 0.304253262 0.82568652  0.320559921
#>    34   19 0.332966393 0.81894979 0.334520205 0.74590880  0.324442251
#>    78   66 0.508496468 0.81824680 0.450348281 0.87633355  0.469306728
#>    36   21 0.074923124 0.61521099 0.000000000 0.67951561  0.013406778
#>    34   80 0.332966393 0.81894979 0.287373143 0.77539185  0.312208926
#>    30   33 0.543791747 0.99239376 0.635000259 0.89187060  0.548783193
#>    85   27 0.265040030 0.80498824 0.244312197 0.69055346  0.220256256
#>    15   12 0.304253262 0.82568652 0.315958737 0.87638469  0.306176432
#>     2   79 0.332444185 0.85018207 0.336872915 0.82570746  0.320556862
#>     7   69 0.715310110 0.84107060 0.755396654 0.92873593  0.853349011
#>    88   39 0.990105321 0.85227660 0.859283593 0.63357510  0.855244381
#>    88   28 0.990105321 0.85227660 0.971809774 0.49540176  1.118130743
#>    88   53 0.990105321 0.85227660 0.997785932 0.34052160  2.152375663
#>    88   44 0.990105321 0.85227660 1.000000000 0.27641825 15.745113063
#>    88   29 0.990105321 0.85227660 0.975270847 0.90741244  0.869944825
#>    15   71 0.304253262 0.82568652 0.266613992 0.86503245  0.295010487
#>    15   85 0.304253262 0.82568652 0.265040030 0.80498824  0.274380842
#>    30   24 0.543791747 0.99239376 0.511473064 0.82640792  0.528215843
#>    55   54 0.781936001 0.99109489 0.981361986 0.93239637  0.876630550
#>    52   73 0.206217915 0.68761442 0.159037593 0.82903802  0.200042027
#>   100   71 0.151144603 0.91913196 0.266613992 0.86503245  0.202388659
#>    67   12 0.318970681 0.90934529 0.315958737 0.87638469  0.285517439
#>    67    2 0.318970681 0.90934529 0.332444185 0.85018207  0.364402270
#>    55   29 0.781936001 0.99109489 0.975270847 0.90741244  0.865398996
#>    54   10 0.981361986 0.93239637 0.988221339 0.87088491  1.031750132
#>    30   66 0.543791747 0.99239376 0.450348281 0.87633355  0.412252645
#>    14   55 0.549674654 1.00000000 0.781936001 0.99109489  0.669316775
#>    55   69 0.781936001 0.99109489 0.755396654 0.92873593  0.790225457
#>    75   14 0.244334309 0.97161912 0.549674654 1.00000000  0.384366597
#>    33   14 0.635000259 0.89187060 0.549674654 1.00000000  0.652478959
#>    90    5 0.048759188 0.83308685 0.067131193 0.64958482  0.068261366
#>   100   73 0.151144603 0.91913196 0.159037593 0.82903802  0.105317848
#>    75   30 0.244334309 0.97161912 0.543791747 0.99239376  0.388093552
#>    71   75 0.266613992 0.86503245 0.244334309 0.97161912  0.264790891
#>    87  100 0.135717350 0.93626803 0.151144603 0.91913196  0.093996786
#>    87   35 0.135717350 0.93626803 0.108310791 0.97121336  0.116436689
#>   100   18 0.151144603 0.91913196 0.135601349 0.96263417  0.190737983
#>    87   90 0.135717350 0.93626803 0.048759188 0.83308685  0.064143708
#>    75   51 0.244334309 0.97161912 0.022316312 0.98955543  0.256162444
#>    51   90 0.022316312 0.98955543 0.048759188 0.83308685 -0.290387462
#>    75   35 0.244334309 0.97161912 0.108310791 0.97121336  0.174765089
#>     1   59 0.615721078 0.00000000 0.709340429 0.01973176  0.657059282
#>    77   47 0.844983332 0.45061350 0.792817768 0.33915347  0.856039989
#>    82   56 0.469192487 0.41401921 0.329129774 0.22626872  0.393335404
#>    64   16 0.621886674 0.33716991 0.727182782 0.55113040  0.642161493
#>    77   13 0.844983332 0.45061350 0.893880853 0.47357700  0.857939495
#>    13   28 0.893880853 0.47357700 0.971809774 0.49540176  0.915077125
#>     6    1 0.077456580 0.01715316 0.615721078 0.00000000  0.265442390
#>    41   62 0.552962856 0.66005668 0.546057422 0.58358097  0.580380824
#>    93   91 0.606648308 0.66080265 0.673160356 0.70695637  0.609397437
#>    24   78 0.511473064 0.82640792 0.508496468 0.81824680  0.601036688
#>    32   65 0.156832115 0.51313304 0.144998001 0.59087816  0.122741739
#>    27   65 0.244312197 0.69055346 0.144998001 0.59087816  0.244521891
#>    23   36 0.076129187 0.56133640 0.074923124 0.61521099  0.011161332
#>    85   52 0.265040030 0.80498824 0.206217915 0.68761442  0.220256256
#>    15    2 0.304253262 0.82568652 0.332444185 0.85018207  0.306176432
#>     7   88 0.715310110 0.84107060 0.990105321 0.85227660  0.853349011
#>    57   33 0.751155912 0.96380903 0.635000259 0.89187060  0.686651358
#>    88   10 0.990105321 0.85227660 0.988221339 0.87088491  0.869944825
#>    21   83 0.000000000 0.67951561 0.011689065 0.30191898 -3.940200476
#>    67   71 0.318970681 0.90934529 0.266613992 0.86503245  0.285517439
#>    51   21 0.022316312 0.98955543 0.000000000 0.67951561 -0.290387462
#>           my1          mx2           my2 bp1 bp2
#>   0.035825839  0.681334234   0.074647203   0   0
#>   0.043579513  0.657059282   0.035825839   0   0
#>   0.128356438  0.681334234   0.074647203   0   0
#>   0.220767248  0.750989267   0.090536635   0   0
#>   0.090536635  0.866324497   0.163718220   0   0
#>   0.075760160  4.527145810  -9.867325530   0   1
#>   0.165533240  0.649432910   0.222002691   0   0
#>   0.218407264  0.673809006   0.220767248   0   0
#>   0.063669889  0.647667292   0.128356438   0   0
#>   0.098187808  0.505608273   0.063669889   0   0
#>   0.222002691  0.629816973   0.261837207   0   0
#>   0.245652329  0.751105304   0.261462258   0   0
#>   0.160364761  0.486419638   0.098187808   0   0
#>   0.311705453  0.629816973   0.261837207   0   0
#>   0.261837207  0.689645006   0.436713954   0   0
#>   0.261462258  0.856039989   0.377501471   0   0
#>   0.163718220  0.929356885   0.136841143   0   0
#>   0.435525558  0.774568302   0.574366492   0   0
#>   0.335999033  0.550095782   0.305354455   0   0
#>   0.305354455  0.469663751   0.165533240   0   0
#>   0.424801937  0.642161493   0.460081955   0   0
#>  -0.556741285  0.509385667   0.043579513   0   0
#>   0.025982508  0.486419638   0.098187808   0   0
#>   0.169628835  0.461213753   0.160364761   0   0
#>   0.201022629  0.881621240   0.373704349   0   0
#>   0.046011944  0.309245773   0.043538221   0   0
#>   0.063617841  0.264152869   0.075651137   0   0
#>   0.073798271  0.309245773   0.043538221   0   0
#>   0.150565974  0.363424628   0.144394287   0   0
#>   0.144394287  0.344033167   0.063617841   0   0
#>   0.306859827  0.929356885   0.136841143   0   0
#>   0.324489984  0.443969236   0.169628835   0   0
#>   0.431608194  0.751105304   0.261462258   0   0
#>  -0.031894950  0.309245773   0.043538221   0   0
#>   0.124785251  0.320423176   0.150565974   0   0
#>   0.338783368  0.387723528   0.328676469   0   0
#>   0.084923785  0.264152869   0.075651137   0   0
#>   0.093189398  0.261246687   0.084923785   0   0
#>   0.460081955  0.631172570   0.508012765   0   0
#>   0.382284783  0.952276083   0.306859827   0   0
#>   0.328676469  0.478982608   0.471114718   0   0
#>   0.471114718  0.437045238   0.530749405   0   0
#>   0.425728567  0.478982608   0.471114718   0   0
#>   0.486567105  0.892211564   0.548416768   0   0
#>   0.382117113  0.881621240   0.373704349   0   0
#>   0.435416741  0.780632605   0.464348040   0   0
#>   0.464348040  0.841576690   0.565715481   0   0
#>   0.567807464  0.793331061   0.739024843   0   0
#>   0.227449078  0.202576842   0.159890003   0   0
#>   0.266791954  0.320423176   0.150565974   0   0
#>   0.392157032  0.160383537   0.294092098   0   0
#>   0.082377579  0.162374485   0.019030009   0   0
#>   0.115766802  0.078140335   0.135901887   0   0
#>   0.159890003  0.167362716   0.115766802   0   0
#>   0.019030009  0.158641662   0.008876621   0   0
#>   0.294092098  0.128063133   0.318108996   0   0
#>   0.566534594  0.443275117   0.622600561   0   0
#>   0.346657543  0.366723682   0.338783368   0   0
#>   0.547933657  1.118130743   0.666806861   0   0
#>   0.548416768  0.861273279   0.567360782   0   0
#>   0.412623910  0.914121243   0.382284783   0   0
#>   0.486126853  0.219104579   0.458003739   0   0
#>   0.523160288  0.598021062   0.605190171   0   0
#>   0.265700567  0.255700652   0.266791954   0   0
#>   0.508012765  0.697471290   0.589368142   0   0
#>  -2.537790184  0.418260176  -0.556741285   0   0
#>  -0.169874287  0.245159894  -0.031894950   0   0
#>   0.123953986  0.078140335   0.135901887   0   0
#>   0.451525463  0.105881491   0.392668459   0   0
#>   0.458003739  0.227766178   0.563703630   0   0
#>   0.485956280  0.362928035   0.346657543   0   0
#>   0.589368142  0.652124187   0.656399049   0   0
#>   0.719311348  0.826281381   0.568552193   0   0
#>   0.568552193  0.774568302   0.574366492   0   0
#>   0.608783810  0.697471290   0.589368142   0   0
#>   0.530749405  0.367525364   0.485956280   0   0
#>   0.619031331  0.598021062   0.605190171   0   0
#>   0.696692460  0.747777144   0.608783810   0   0
#>   0.567360782  0.911995455   0.750522074   0   0
#>   0.630419446  0.482286434   0.674097422   0   0
#>   0.727842949  0.610956538   0.790514187   0   0
#>   0.663536795  0.652124187   0.656399049   0   0
#>   0.789118053  0.499601010   0.755168204   0   0
#>   0.674097422  0.579392429   0.690162894   0   0
#>   0.392668459  0.080350110   0.410213230   0   0
#>   0.318108996 -0.055238480   0.136484353   0   0
#>   0.594324252  0.367109850   0.486126853   0   0
#>   0.657780280  0.443275117   0.622600561   0   0
#>   0.136484353 -3.389071695  -0.633474381   0   1
#>   0.791426768  0.548783193   0.905282620   0   0
#>   0.547717145  0.105061105   0.588934882   0   0
#>   0.591029676  0.227766178   0.563703630   0   0
#>   0.653585043  0.258243669   0.594324252   0   0
#>   0.783688800  0.376946796   0.657780280   0   0
#>   0.410213230 -3.940200476   0.368561607   0   0
#>   0.644485349  0.013406778   0.619336397   0   0
#>   0.603649191  0.220256256   0.754005398   0   0
#>   0.752837199  0.286941771   0.722263506   0   0
#>   0.755168204  0.482286434   0.674097422   0   0
#>   0.806687175  0.395752986   0.787161684   0   0
#>   0.790514187  0.629533990   0.794347950   0   0
#>   0.479852495  0.011161332   0.586832796   0   0
#>   0.586832796 -0.005151834   0.597713236   0   0
#>   0.742445541  0.855244381   0.784469135   0   0
#>   0.662899028  0.118033281   0.736776864   0   0
#>   0.722263506  0.329085700   0.653585043   0   0
#>   0.830629794  0.320556862   0.835393036   0   0
#>   0.782231432  0.286941771   0.722263506   0   0
#>   0.837163834  0.528215843   0.909287236   0   0
#>   0.619336397  0.068261366   0.742368675   0   0
#>   0.736776864  0.100761504   0.745622551   0   0
#>   0.795036379  0.291735284   0.801907834   0   0
#>   0.794347950  0.697704511   0.902118883   0   0
#>   0.801907834  0.226705679   0.752837199   0   0
#>   0.848792153  0.413389547   0.806687175   0   0
#>   0.804988315  0.499601010   0.755168204   0   0
#>   0.851942881  0.364402270   0.888575827   0   0
#>   0.830947645  0.858382738   0.846399772   0   0
#>   0.613785258  0.952971940   0.412623910   0   0
#>   0.817790171  0.952276083   0.306859827   0   0
#>   0.938216758  0.697704511   0.902118883   0   0
#>   0.849510589  1.031750132   0.906877141   0   0
#>   0.854520937  0.285517439   0.895784334   0   0
#>   0.834786119  0.216415872   0.836305577   0   0
#>   0.893210316  0.394604760   0.848792153   0   0
#>   0.953505398  0.384366597   1.121776295   0   0
#>   0.597713236 -3.940200476   0.368561607   0   0
#>   0.836305577  0.202388659   0.878228659   0   0
#>   0.846399772  0.865398996   0.918746958   0   0
#>   0.944695657  1.031750132   0.906877141   0   0
#>   0.764135797  0.216415872   0.836305577   0   0
#>   1.002652602  0.384752677   0.893210316   0   0
#>   1.087132322  0.790225457   0.950740060   0   0
#>   0.950740060  0.686651358   0.938216758   0   0
#>   0.993393352  0.669316775   1.087132322   0   0
#>   0.869724435  0.100761504   0.745622551   0   0
#>   0.910377881  0.190737983   0.957806489   0   0
#>   1.068053845  0.412252645   1.002652602   0   0
#>   0.920273248  0.388093552   1.068053845   0   0
#>   0.883195411  0.105317848   0.869724435   0   0
#>   0.949592278  0.116436689   0.949366527   0   0
#>   0.949366527  0.175824798   1.138281020   0   0
#>   0.957806489  0.175824798   1.138281020   0   0
#>   0.908354735  0.051145404   0.913958798   0   0
#>   2.501081517  0.218621190   4.396581820   0   1
#>   0.856240429  0.068261366   0.742368675   0   0
#>   1.493534320  0.051145404   0.913958798   0   0
#>   0.074647203  0.738276689   0.075760160   0   0
#>   0.043579513  0.505608273   0.063669889   0   0
#>   0.128356438  0.656219522   0.218407264   0   0
#>   0.220767248  0.750462591   0.245652329   0   0
#>   0.075760160  0.750989267   0.090536635   0   0
#>   0.218407264  0.649432910   0.222002691   0   0
#>   0.160364761  0.469663751   0.165533240   0   0
#>   0.311705453  0.580432132   0.424801937   0   0
#>   0.163718220  0.833505508   0.201022629   0   0
#>   0.435525558  0.716291245   0.431608194   0   0
#>   0.335999033  0.551711264   0.425728567   0   0
#>   0.305354455  0.567985210   0.311705453   0   0
#>   0.424801937  0.551711264   0.425728567   0   0
#>  -0.556741285  0.352012969   0.025982508   0   0
#>   0.025982508  0.342771982   0.046011944   0   0
#>   0.377501471  0.881621240   0.373704349   0   0
#>   0.201022629  0.750462591   0.245652329   0   0
#>   0.063617841  0.342771982   0.046011944   0   0
#>   0.073798271  0.264152869   0.075651137   0   0
#>   0.144394287  0.443969236   0.169628835   0   0
#>   0.136841143  3.383108274  -2.092852008   0   1
#>   0.324489984  0.437662966   0.335999033   0   0
#>  -0.031894950  0.264448793   0.073798271   0   0
#>   0.124785251  0.202576842   0.159890003   0   0
#>   0.093189398  0.270910137   0.124785251   0   0
#>   0.486567105  0.780632605   0.464348040   0   0
#>   0.382117113  0.914121243   0.382284783   0   0
#>   0.435416741  0.728942895   0.435525558   0   0
#>   0.567807464  0.841576690   0.565715481   0   0
#>   0.227449078  0.190962054   0.265700567   0   0
#>   0.266791954  0.366723682   0.338783368   0   0
#>   0.392157032  0.362928035   0.346657543   0   0
#>   0.436713954  0.716291245   0.431608194   0   0
#>   0.082377579  0.260563447   0.093189398   0   0
#>   0.115766802  0.175938337   0.082377579   0   0
#>   0.019030009  0.261246687   0.084923785   0   0
#>   0.294092098  0.190962054   0.265700567   0   0
#>   0.566534594  0.437045238   0.530749405   0   0
#>   0.547933657  0.892211564   0.548416768   0   0
#>   0.486126853  0.367525364   0.485956280   0   0
#>   0.523160288  0.448999409   0.566534594   0   0
#>   0.135901887  0.185834063   0.227449078   0   0
#>   0.508012765  0.618663374   0.523160288   0   0
#>  -2.537790184  0.198378508  -0.169874287   0   0
#>  -0.169874287  0.158641662   0.008876621   0   0
#>   0.008876621  0.014175746   0.123953986   0   0
#>   0.123953986 -0.055238480   0.136484353   0   0
#>   0.451525463  0.282456558   0.392157032   0   0
#>   0.458003739  0.215934437   0.451525463   0   0
#>   0.719311348  0.793331061   0.739024843   0   0
#>   0.568552193  0.841576690   0.565715481   0   0
#>   0.605190171  0.652124187   0.656399049   0   0
#>   0.608783810  0.774568302   0.574366492   0   0
#>   0.619031331  0.579392429   0.690162894   0   0
#>   0.739024843  0.794679711   0.742445541   0   0
#>   0.696692460  0.780817592   0.719311348   0   0
#>   0.567360782  0.859518294   0.567807464   0   0
#>   0.630419446  0.443275117   0.622600561   0   0
#>   0.727842949  0.579392429   0.690162894   0   0
#>   0.663536795  0.765407852   0.696692460   0   0
#>   0.789118053  0.609120597   0.791426768   0   0
#>   0.392668459  0.128063133   0.318108996   0   0
#>   0.547717145  0.082206601   0.479852495   0   0
#>   0.591029676  0.258243669   0.594324252   0   0
#>   0.653585043  0.376946796   0.657780280   0   0
#>   0.783688800  0.395752986   0.787161684   0   0
#>   0.644485349  0.105061105   0.588934882   0   0
#>   0.603649191  0.138233231   0.662899028   0   0
#>   0.750522074  1.118130743   0.666806861   0   0
#>   0.806687175  0.437165204   0.804988315   0   0
#>   0.787161684  0.320559921   0.830629794   0   0
#>   0.790514187  0.609120597   0.791426768   0   0
#>   0.479852495  0.080350110   0.410213230   0   0
#>   0.742445541  0.629533990   0.794347950   0   0
#>   0.662899028  0.124350479   0.644485349   0   0
#>   0.830629794  0.312208926   0.795036379   0   0
#>   0.782231432  0.392949611   0.783688800   0   0
#>   0.837163834  0.437165204   0.804988315   0   0
#>   0.619336397 -0.005151834   0.597713236   0   0
#>   0.795036379  0.324442251   0.782231432   0   0
#>   0.905282620  0.601930706   0.953505398   0   0
#>   0.754005398  0.226705679   0.752837199   0   0
#>   0.851942881  0.295010487   0.854520937   0   0
#>   0.835393036  0.394604760   0.848792153   0   0
#>   0.830947645  0.697704511   0.902118883   0   0
#>   0.784469135  0.911995455   0.750522074   0   0
#>   0.666806861  2.152375663   0.613785258   0   0
#>   0.613785258 15.745113063   0.817790171   0   0
#>   0.817790171 17.481138679   0.847619411   0   1
#>   0.849510589  0.858382738   0.846399772   0   0
#>   0.854520937  0.274380842   0.834786119   0   0
#>   0.834786119  0.291735284   0.801907834   0   0
#>   0.909287236  0.548783193   0.905282620   0   0
#>   0.944695657  2.234882002   5.559303175   0   1
#>   0.764135797  0.118033281   0.736776864   0   0
#>   0.878228659  0.217451156   0.910377881   0   0
#>   0.895784334  0.364402270   0.888575827   0   0
#>   0.888575827  0.384752677   0.893210316   0   0
#>   0.918746958  0.876630550   0.944695657   0   0
#>   0.906877141 17.089206454   2.697498995   0   1
#>   1.002652602  0.528215843   0.909287236   0   0
#>   1.087132322  0.834151032   5.386307123   0   1
#>   0.950740060  0.865398996   0.918746958   0   0
#>   1.121776295  0.256162444   2.501081517   0   0
#>   0.993393352  0.601930706   0.953505398   0   0
#>   0.742368675  0.100761504   0.745622551   0   0
#>   0.869724435  0.202388659   0.878228659   0   0
#>   1.068053845  0.384366597   1.121776295   0   0
#>   0.920273248  0.217451156   0.910377881   0   0
#>   0.883195411  0.167748183   0.949592278   0   0
#>   0.949366527  0.064143708   0.908354735   0   0
#>   0.957806489  0.167748183   0.949592278   0   0
#>   0.908354735  0.093996786   0.883195411   0   0
#>   2.501081517  0.174765089   1.493534320   0   0
#>   0.856240429  0.051145404   0.913958798   0   0
#>   1.493534320  0.175824798   1.138281020   0   0
#>   0.035825839  2.812610609 -10.191410036   0   1
#>   0.377501471  0.732294801   0.435416741   0   0
#>   0.324489984  0.387723528   0.328676469   0   0
#>   0.460081955  0.689645006   0.436713954   0   0
#>   0.486567105  0.906991837   0.382117113   0   0
#>   0.547933657  0.952971940   0.412623910   0   0
#>  -2.537790184  0.206298219  -4.393728100   0   1
#>   0.619031331  0.454260705   0.630419446   0   0
#>   0.727842949  0.654020450   0.663536795   0   0
#>   0.789118053  0.469306728   0.837163834   0   0
#>   0.547717145  0.227766178   0.563703630   0   0
#>   0.591029676  0.231856491   0.603649191   0   0
#>   0.586832796  0.105061105   0.588934882   0   0
#>   0.754005398  0.200042027   0.764135797   0   0
#>   0.851942881  0.320556862   0.835393036   0   0
#>   0.830947645  0.855244381   0.784469135   0   0
#>   0.938216758  0.652478959   0.993393352   0   0
#>   0.849510589 54.064117507   6.235106952   0   1
#>   0.368561607 -6.585993684   0.286657145   0   1
#>   0.895784334  0.264790891   0.920273248   0   0
#>   0.856240429 -3.499155199   1.087203866   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.615721078 0.00000000
#>   [2,] 0.332444185 0.85018207
#>   [3,] 0.682229523 0.31652569
#>   [4,] 0.707173183 0.13061507
#>   [5,] 0.067131193 0.64958482
#>   [6,] 0.077456580 0.01715316
#>   [7,] 0.715310110 0.84107060
#>   [8,] 0.132421946 0.08615563
#>   [9,] 0.431204848 0.43835760
#>  [10,] 0.988221339 0.87088491
#>  [11,] 0.381977327 0.05758785
#>  [12,] 0.315958737 0.87638469
#>  [13,] 0.893880853 0.47357700
#>  [14,] 0.549674654 1.00000000
#>  [15,] 0.304253262 0.82568652
#>  [16,] 0.727182782 0.55113040
#>  [17,] 0.693789517 0.63765211
#>  [18,] 0.135601349 0.96263417
#>  [19,] 0.334520205 0.74590880
#>  [20,] 0.394907875 0.03351826
#>  [21,] 0.000000000 0.67951561
#>  [22,] 0.797160985 0.05319628
#>  [23,] 0.076129187 0.56133640
#>  [24,] 0.511473064 0.82640792
#>  [25,] 0.824223349 0.31075329
#>  [26,] 0.300846034 0.44265988
#>  [27,] 0.244312197 0.69055346
#>  [28,] 0.971809774 0.49540176
#>  [29,] 0.975270847 0.90741244
#>  [30,] 0.543791747 0.99239376
#>  [31,] 0.544674827 0.71015276
#>  [32,] 0.156832115 0.51313304
#>  [33,] 0.635000259 0.89187060
#>  [34,] 0.332966393 0.81894979
#>  [35,] 0.108310791 0.97121336
#>  [36,] 0.074923124 0.61521099
#>  [37,] 0.009460254 0.51706178
#>  [38,] 0.234362040 0.36816479
#>  [39,] 0.859283593 0.63357510
#>  [40,] 0.289521198 0.50225981
#>  [41,] 0.552962856 0.66005668
#>  [42,] 0.815455127 0.61663240
#>  [43,] 0.554513527 0.25363377
#>  [44,] 1.000000000 0.27641825
#>  [45,] 0.305616941 0.06305787
#>  [46,] 0.248661331 0.35036477
#>  [47,] 0.792817768 0.33915347
#>  [48,] 0.897040431 0.62311924
#>  [49,] 0.307486769 0.07547030
#>  [50,] 0.589650410 0.14177634
#>  [51,] 0.022316312 0.98955543
#>  [52,] 0.206217915 0.68761442
#>  [53,] 0.997785932 0.34052160
#>  [54,] 0.981361986 0.93239637
#>  [55,] 0.781936001 0.99109489
#>  [56,] 0.329129774 0.22626872
#>  [57,] 0.751155912 0.96380903
#>  [58,] 0.898194888 0.29014199
#>  [59,] 0.709340429 0.01973176
#>  [60,] 0.827870384 0.51015465
#>  [61,] 0.289782762 0.03961695
#>  [62,] 0.546057422 0.58358097
#>  [63,] 0.148525646 0.15751418
#>  [64,] 0.621886674 0.33716991
#>  [65,] 0.144998001 0.59087816
#>  [66,] 0.450348281 0.87633355
#>  [67,] 0.318970681 0.90934529
#>  [68,] 0.440070369 0.73249351
#>  [69,] 0.755396654 0.92873593
#>  [70,] 0.654023440 0.61074995
#>  [71,] 0.266613992 0.86503245
#>  [72,] 0.736870487 0.14821497
#>  [73,] 0.159037593 0.82903802
#>  [74,] 0.503165228 0.28317184
#>  [75,] 0.244334309 0.97161912
#>  [76,] 0.212248233 0.10665805
#>  [77,] 0.844983332 0.45061350
#>  [78,] 0.508496468 0.81824680
#>  [79,] 0.336872915 0.82570746
#>  [80,] 0.287373143 0.77539185
#>  [81,] 0.619700138 0.07578372
#>  [82,] 0.469192487 0.41401921
#>  [83,] 0.011689065 0.30191898
#>  [84,] 0.798933982 0.52755178
#>  [85,] 0.265040030 0.80498824
#>  [86,] 0.217186468 0.06800602
#>  [87,] 0.135717350 0.93626803
#>  [88,] 0.990105321 0.85227660
#>  [89,] 0.350499101 0.56361462
#>  [90,] 0.048759188 0.83308685
#>  [91,] 0.673160356 0.70695637
#>  [92,] 0.525967976 0.50499862
#>  [93,] 0.606648308 0.66080265
#>  [94,] 0.520911802 0.34828278
#>  [95,] 0.251482332 0.18302935
#>  [96,] 0.308521147 0.02369735
#>  [97,] 0.228673995 0.05076923
#>  [98,] 0.110804258 0.20188864
#>  [99,] 0.808721339 0.61460172
#> [100,] 0.151144603 0.91913196
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,59,81), [2,3,0]
#> 2: (4,81,59), [1,7,4]
#> 3: (20,1,81), [1,10,23]
#> 4: (50,81,4), [2,9,10]
#> 5: (3,4,72), [6,13,9]
#> 6: (4,22,72), [18,5,7]
#> 7: (4,59,22), [0,6,2]
#> 8: (74,50,43), [12,21,14]
#> 9: (50,4,3), [5,12,4]
#> 10: (20,81,50), [4,11,3]
#> 11: (11,20,50), [10,14,24]
#> 12: (50,3,43), [16,8,9]
#> 13: (72,25,3), [17,5,27]
#> 14: (74,11,50), [11,8,25]
#> 15: (94,43,64), [16,22,21]
#> 16: (43,3,64), [58,15,12]
#> 17: (3,25,47), [26,37,13]
#> 18: (72,22,58), [35,27,6]
#> 19: (47,84,16), [74,37,52]
#> 20: (82,74,94), [21,49,36]
#> 21: (94,74,43), [8,15,20]
#> 22: (94,64,92), [44,49,15]
#> 23: (96,1,20), [3,24,75]
#> 24: (96,20,11), [11,28,23]
#> 25: (56,11,74), [14,36,33]
#> 26: (77,47,25), [17,45,52]
#> 27: (72,58,25), [45,13,18]
#> 28: (11,45,96), [30,24,29]
#> 29: (11,49,45), [39,28,33]
#> 30: (61,96,45), [28,31,38]
#> 31: (97,61,45), [30,39,38]
#> 32: (95,49,56), [33,56,40]
#> 33: (56,49,11), [29,25,32]
#> 34: (53,58,44), [35,145,46]
#> 35: (44,58,22), [18,0,34]
#> 36: (82,56,74), [25,20,47]
#> 37: (16,3,47), [17,19,58]
#> 38: (97,96,61), [30,31,76]
#> 39: (97,45,49), [29,42,31]
#> 40: (76,49,95), [32,61,43]
#> 41: (46,56,9), [47,65,56]
#> 42: (86,97,49), [39,43,62]
#> 43: (76,86,49), [42,40,59]
#> 44: (64,16,92), [73,22,58]
#> 45: (77,25,58), [27,51,26]
#> 46: (13,58,53), [34,68,51]
#> 47: (56,82,9), [48,41,36]
#> 48: (82,92,9), [88,47,49]
#> 49: (94,92,82), [48,20,22]
#> 50: (77,13,60), [67,53,51]
#> 51: (13,77,58), [45,46,50]
#> 52: (47,77,84), [53,19,26]
#> 53: (77,60,84), [94,52,50]
#> 54: (60,39,42), [90,94,92]
#> 55: (98,63,95), [61,72,71]
#> 56: (46,95,56), [32,41,72]
#> 57: (26,38,46), [63,65,79]
#> 58: (16,64,3), [16,37,44]
#> 59: (76,8,86), [62,43,60]
#> 60: (76,63,8), [71,59,61]
#> 61: (95,63,76), [60,40,55]
#> 62: (86,8,97), [77,42,59]
#> 63: (46,38,98), [101,72,57]
#> 64: (92,62,89), [105,88,70]
#> 65: (26,46,9), [41,81,57]
#> 66: (13,28,48), [143,67,68]
#> 67: (13,48,60), [92,50,66]
#> 68: (28,13,53), [46,144,66]
#> 69: (89,40,26), [80,81,102]
#> 70: (92,70,62), [86,64,73]
#> 71: (98,8,63), [60,55,78]
#> 72: (98,95,46), [56,63,55]
#> 73: (92,16,70), [82,70,44]
#> 74: (99,16,84), [19,84,87]
#> 75: (6,1,96), [23,76,0]
#> 76: (6,96,97), [38,77,75]
#> 77: (6,97,8), [62,78,76]
#> 78: (6,8,98), [71,104,77]
#> 79: (26,32,38), [100,57,80]
#> 80: (26,40,32), [108,79,69]
#> 81: (89,26,9), [65,88,69]
#> 82: (16,17,70), [85,73,87]
#> 83: (91,99,42), [84,90,91]
#> 84: (42,99,84), [74,94,83]
#> 85: (93,70,17), [82,96,86]
#> 86: (93,62,70), [70,85,89]
#> 87: (99,17,16), [82,74,91]
#> 88: (92,89,9), [81,48,64]
#> 89: (41,62,93), [86,99,93]
#> 90: (91,42,39), [54,123,83]
#> 91: (91,17,99), [87,83,96]
#> 92: (60,48,39), [117,54,67]
#> 93: (62,41,68), [98,105,89]
#> 94: (42,84,60), [53,54,84]
#> 95: (93,91,31), [120,99,96]
#> 96: (91,93,17), [85,91,95]
#> 97: (24,78,31), [116,106,129]
#> 98: (68,41,31), [99,116,93]
#> 99: (93,31,41), [98,89,95]
#> 100: (38,32,83), [112,101,79]
#> 101: (38,83,98), [104,63,100]
#> 102: (27,40,89), [69,110,109]
#> 103: (19,89,68), [105,111,110]
#> 104: (98,83,6), [0,78,101]
#> 105: (62,68,89), [103,64,93]
#> 106: (31,33,24), [134,97,120]
#> 107: (32,65,23), [124,121,108]
#> 108: (65,32,40), [80,109,107]
#> 109: (27,65,40), [108,102,114]
#> 110: (19,27,89), [102,103,126]
#> 111: (34,19,68), [103,119,128]
#> 112: (32,37,83), [154,100,121]
#> 113: (65,5,36), [130,124,125]
#> 114: (65,27,52), [135,125,109]
#> 115: (85,27,80), [126,136,135]
#> 116: (78,68,31), [98,97,138]
#> 117: (88,39,48), [92,143,142]
#> 118: (66,79,68), [119,138,137]
#> 119: (34,68,79), [118,127,111]
#> 120: (31,91,33), [133,106,95]
#> 121: (32,23,37), [122,112,107]
#> 122: (23,36,37), [155,121,124]
#> 123: (91,39,7), [142,133,90]
#> 124: (36,23,65), [107,113,122]
#> 125: (65,52,5), [131,113,114]
#> 126: (80,27,19), [110,128,115]
#> 127: (34,79,15), [140,132,119]
#> 128: (34,80,19), [126,111,132]
#> 129: (78,24,66), [158,138,97]
#> 130: (36,5,21), [171,155,113]
#> 131: (52,73,5), [153,125,160]
#> 132: (34,15,80), [136,128,127]
#> 133: (91,7,33), [147,120,123]
#> 134: (30,24,33), [106,152,158]
#> 135: (85,52,27), [114,115,160]
#> 136: (15,85,80), [115,132,150]
#> 137: (2,79,66), [118,151,140]
#> 138: (66,68,78), [116,129,118]
#> 139: (15,2,12), [163,149,140]
#> 140: (2,15,79), [127,137,139]
#> 141: (7,88,69), [157,147,142]
#> 142: (88,7,39), [123,117,141]
#> 143: (88,48,28), [66,144,117]
#> 144: (88,28,53), [68,145,143]
#> 145: (88,53,44), [34,0,144]
#> 146: (57,33,69), [147,168,170]
#> 147: (7,69,33), [146,133,141]
#> 148: (88,10,29), [165,157,0]
#> 149: (15,12,71), [162,150,139]
#> 150: (15,71,85), [156,136,149]
#> 151: (67,2,66), [137,166,163]
#> 152: (33,14,30), [169,134,170]
#> 153: (90,5,73), [131,172,171]
#> 154: (21,83,37), [112,155,0]
#> 155: (36,21,37), [154,122,130]
#> 156: (85,71,73), [161,160,150]
#> 157: (88,29,69), [164,141,148]
#> 158: (30,66,24), [129,134,166]
#> 159: (55,29,54), [165,0,164]
#> 160: (52,85,73), [156,131,135]
#> 161: (100,73,71), [156,173,172]
#> 162: (67,71,12), [149,163,175]
#> 163: (67,12,2), [139,151,162]
#> 164: (55,69,29), [157,159,168]
#> 165: (54,29,10), [148,0,159]
#> 166: (30,67,66), [151,158,174]
#> 167: (14,57,55), [168,0,170]
#> 168: (55,57,69), [146,164,167]
#> 169: (75,30,14), [152,182,174]
#> 170: (33,57,14), [167,152,146]
#> 171: (90,21,5), [130,153,183]
#> 172: (100,90,73), [153,161,176]
#> 173: (71,75,100), [179,161,175]
#> 174: (75,67,30), [166,169,175]
#> 175: (71,67,75), [174,173,162]
#> 176: (87,90,100), [172,177,181]
#> 177: (100,18,87), [178,176,179]
#> 178: (87,18,35), [180,181,177]
#> 179: (100,75,18), [180,177,173]
#> 180: (75,35,18), [178,179,185]
#> 181: (87,35,90), [184,176,178]
#> 182: (75,14,51), [0,185,169]
#> 183: (51,21,90), [171,184,0]
#> 184: (51,90,35), [181,185,183]
#> 185: (75,51,35), [184,180,182]
#> boundary nodes:  1 59 22 44 88 10 54 55 14 51 21 83 6 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.105722298 0.075888109 0.103237641 0.034520810 0.112532201 0.093980446 
#>                                                                         
#> 0.117246260 0.187576501 0.072512153 0.222810382 0.142361401 0.142111108 
#>                                                                         
#> 0.224088846 0.107319274 0.063776510 0.042342424 0.257587113 0.075526062 
#>                                                                         
#> 0.067486112 0.059238037 0.193305128 0.223342662 0.027322959 0.256075461 
#>                                                                         
#> 0.076789416 0.039467516 0.012552478 0.028287769 0.152343633 0.076606953 
#>                                                                         
#> 0.102725954 0.183101895 0.112879478 0.024587816 0.121266011 0.235374197 
#>                                                                         
#> 0.082592954 0.090608277 0.206435317 0.111608609 0.045115603 0.115849367 
#>                                                                         
#> 0.107241354 0.075467254 0.169063820 0.089666283 0.033763523 0.046989250 
#>                                                                         
#> 0.106071217 0.088875016 0.022832222 0.086685827 0.073153073 0.081528608 
#>                                                                         
#> 0.102550739 0.207157633 0.196574953 0.202644696 0.147993948 0.132459399 
#>                                                                         
#> 0.168819445 0.060666327 0.111332002 0.167359203 0.094375737 0.308112580 
#>                                                                         
#> 0.084311651 0.117734675 0.164397938 0.133133842 0.130429790 0.048011116 
#>                                                                         
#> 0.007033318 0.087598420 0.092742204 0.149005838 0.098155201 0.117220496 
#>                                                                         
#> 0.039177848 0.134133546 0.128525282 0.090163949 0.113987700 0.050777053 
#>                                                                         
#> 0.256276969 0.140818643 0.086502705 0.191162473 0.292261757 0.139801043 
#>                                                                         
#> 0.074937500 0.169529249 0.165496608 0.106399282 0.215154347 0.035245906 
#>                                                                         
#> 0.038207490 0.095140939 0.106963553 0.139063134 0.188810667 0.080031173 
#>                                                                         
#> 0.117977352 0.252552462 0.144192118 0.105838078 0.032619660 0.055606616 
#>                                                                         
#> 0.078922801 0.073501358 0.201619046 0.053051780 0.095027959 0.037077137 
#>                                                                         
#> 0.124256428 0.109707605 0.030957183 0.246848508 0.157043380 0.064141576 
#>                                                                         
#> 0.125914035 0.038755329 0.050633755 0.060064830 0.120769516 0.009615797 
#>                                                                         
#> 0.162729052 0.113438442 0.220905753 0.025715726 0.108696406 0.135461612 
#>                                                                         
#> 0.041133101 0.035328550 0.204705847 0.110352706 0.106954302 0.239669688 
#>                                                                         
#> 0.097204001 0.133740551 0.026366389 0.028607293 0.109103555 0.150417180 
#>                                                                         
#> 0.527461762 0.161125982 0.087928845 0.110904492 0.228731131 0.118051583 
#>                                                                         
#> 0.176957997 0.118707515 0.197757998 0.165747919 0.101584549 0.214869360 
#>                                                                         
#> 0.221905772 0.083643090 0.100436595 0.156797384 0.086943183 0.080897320 
#>                                                                         
#> 0.141392557 0.184524330 0.076556056 0.077918081 0.176765662 0.301615181 
#>                                                                         
#> 0.135185734 0.062118072 0.085859707 0.100215042 0.061951611 0.183485737 
#>                                                                         
#> 0.188497566 0.107199105 0.141936583 0.147901998 0.106026551 0.238872705 
#>                                                                         
#> 0.038966210 0.082417143 0.020714009 0.202607343 0.185000438 0.149575607 
#>                                                                         
#> 0.130749672 0.081109632 0.058240841 0.166076931 0.231157220 0.154908834 
#>                                                                         
#> 0.088218659 0.187721238 0.099848081 0.160332391 0.168541455 0.090599689 
#>                                                                         
#> 0.068917894 0.103330265 0.053690635 0.200066665 0.164030978 0.127355391 
#>                                                                         
#> 0.182779100 0.079222128 0.072309349 0.120903341 0.232318234 0.094002801 
#>                                                                         
#> 0.193644917 0.182993153 0.137644286 0.074179347 0.114480490 0.247334133 
#>                                                                         
#> 0.144206775 0.007805581 0.202928702 0.147424219 0.140581748 0.097517752 
#>                                                                         
#> 0.029492837 0.073057509 0.082190523 0.098734788 0.063055831 0.135734655 
#>                                                                         
#> 0.116296872 0.052031939 0.024872074 0.096395755 0.254842449 0.357343507 
#>                                                                         
#> 0.511812634 0.575943353 0.057096604 0.054450132 0.044340682 0.169102917 
#>                                                                         
#> 0.207885160 0.149085942 0.127514460 0.033097933 0.060678019 0.210668257 
#>                                                                         
#> 0.061892728 0.149002187 0.232432000 0.067771506 0.306656487 0.137740426 
#>                                                                         
#> 0.184419427 0.090439022 0.300177186 0.108890325 0.023057430 0.044410535 
#>                                                                         
#> 0.046195613 0.134937316 0.222741335 0.158687246 0.136024123 0.095676147 
#>                                                                         
#> 0.123063333 0.234238790 0.238466696 0.054021200 0.080927355 0.538537743 
#>                                                                         
#> 0.076786838 0.080956890 0.008686993 0.078640635 0.140707056 0.053888092 
#>                                                                         
#> 0.131288441 0.037346490 0.275023604 0.136628229 0.018703437 0.377777518 
#>                         
#> 0.068591917 0.310841926 
#> 
#> $alpha
#> [1] 0.3029915
#> 
#> attr(,"class")
#> [1] "scree"