Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2         y2         mx1
#>    33   40 0.001873173 0.908574909 0.226823618 0.93053059  0.11483212
#>    40   94 0.226823618 0.930530587 0.408149833 0.95069586  0.26635253
#>    94   28 0.408149833 0.950695864 0.553432315 1.00000000  0.28330374
#>    39   13 0.336163379 0.722655684 0.246381842 0.70344020  0.28465173
#>    13   18 0.246381842 0.703440202 0.214029704 0.54672410  0.24125031
#>    18   31 0.214029704 0.546724104 0.379803082 0.65350974  0.30332009
#>    31   39 0.379803082 0.653509737 0.336163379 0.72265568  0.33519726
#>    82   72 0.749574876 0.430768950 0.826325622 0.44956761  0.76718674
#>    72   44 0.826325622 0.449567612 0.808957076 0.65902462  0.78262684
#>    44   91 0.808957076 0.659024623 0.691650581 0.61570958  0.78162959
#>    91   20 0.691650581 0.615709582 0.633743571 0.57537483  0.67378808
#>    20   81 0.633743571 0.575374832 0.617592474 0.56701190  0.64008109
#>    81   22 0.617592474 0.567011902 0.615968249 0.51206375  0.63784370
#>    22   82 0.615968249 0.512063751 0.749574876 0.43076895  0.67591211
#>    63   19 0.357875639 0.020880277 0.458748142 0.02909651  0.40886011
#>    19   56 0.458748142 0.029096514 0.482643214 0.09860354  0.46537132
#>    56   86 0.482643214 0.098603542 0.363112798 0.13393470  0.42672056
#>    86   63 0.363112798 0.133934702 0.357875639 0.02088028  0.37164855
#>    94   25 0.408149833 0.950695864 0.467919338 0.91784026  0.46039503
#>    25   75 0.467919338 0.917840261 0.504816864 0.91302695  0.49029422
#>    75    7 0.504816864 0.913026953 0.536140570 0.95625271  0.51255701
#>     7   28 0.536140570 0.956252714 0.553432315 1.00000000  0.53876795
#>    79   46 0.938192274 0.925219504 0.880230935 0.89970752  0.94611366
#>    46   48 0.880230935 0.899707515 0.864581824 0.69655025  0.90134103
#>    48   60 0.864581824 0.696550251 0.984024325 0.69975610  0.92327405
#>    60   21 0.984024325 0.699756100 0.982453680 0.41199465  1.76163219
#>    21   71 0.982453680 0.411994653 1.000000000 0.28768139  2.29208431
#>    75   54 0.504816864 0.913026953 0.564440741 0.85962301  0.55530148
#>    54   64 0.564440741 0.859623008 0.616336186 0.94748343  0.57213421
#>    64   28 0.616336186 0.947483426 0.553432315 1.00000000  0.57768764
#>    63   83 0.357875639 0.020880277 0.000000000 0.01668755  0.24381381
#>    83   53 0.000000000 0.016687554 0.784848368 0.00000000  0.18161586
#>    53   76 0.784848368 0.000000000 0.691375983 0.04108218  0.71239860
#>    76   92 0.691375983 0.041082179 0.554546492 0.03876275  0.62213974
#>    92   19 0.554546492 0.038762752 0.458748142 0.02909651  0.52839605
#>    97   80 0.174581138 0.260068306 0.165016429 0.14403433  0.16092692
#>    80   63 0.165016429 0.144034332 0.357875639 0.02088028  0.20961021
#>    86   10 0.363112798 0.133934702 0.223308194 0.20740115  0.27284006
#>    10   97 0.223308194 0.207401146 0.174581138 0.26006831  0.16595452
#>    86   90 0.363112798 0.133934702 0.428306820 0.21067851  0.32544043
#>    90   51 0.428306820 0.210678506 0.344393440 0.40697115  0.32463401
#>    51   97 0.344393440 0.406971147 0.174581138 0.26006831  0.28051984
#>    54   96 0.564440741 0.859623008 0.479997398 0.87108013  0.51862542
#>    96   26 0.479997398 0.871080130 0.462462275 0.85645042  0.49727182
#>    26   45 0.462462275 0.856450418 0.498057006 0.78378495  0.46775747
#>    45   36 0.498057006 0.783784948 0.574952148 0.78407847  0.53660150
#>    36   54 0.574952148 0.784078468 0.564440741 0.85962301  0.54223529
#>    72  100 0.826325622 0.449567612 0.822881876 0.40548598  0.83890545
#>   100   41 0.822881876 0.405485984 0.863748525 0.38427948  0.85169106
#>    41   93 0.863748525 0.384279477 0.921484029 0.40552231  0.88636785
#>    93   37 0.921484029 0.405522308 0.930036397 0.56546045  0.92720633
#>    37   72 0.930036397 0.565460447 0.826325622 0.44956761  0.86721461
#>    39   69 0.336163379 0.722655684 0.210993918 0.75592575  0.28026893
#>    69   35 0.210993918 0.755925750 0.162013222 0.72083072  0.20444533
#>    35   18 0.162013222 0.720830723 0.214029704 0.54672410  0.18804755
#>    52   65 0.999217041 0.274739148 0.973666031 0.27884285  0.96910883
#>    65   95 0.973666031 0.278842853 0.849214684 0.20794312  0.95056684
#>    95   68 0.849214684 0.207943119 0.794296087 0.16386211  0.86303655
#>    68   53 0.794296087 0.163862114 0.784848368 0.00000000  0.82173116
#>    48   74 0.864581824 0.696550251 0.722481668 0.77735017  0.81100185
#>    74   23 0.722481668 0.777350173 0.767863296 0.67330541  0.79295272
#>    23   44 0.767863296 0.673305408 0.808957076 0.65902462  0.79929864
#>    44   37 0.808957076 0.659024623 0.930036397 0.56546045  0.87096053
#>    37   48 0.930036397 0.565460447 0.864581824 0.69655025  0.88303875
#>    50   55 0.104058003 0.512649649 0.067013876 0.46283452  0.10779847
#>    55   70 0.067013876 0.462834518 0.073946410 0.28832545  0.13988497
#>    70   97 0.073946410 0.288325448 0.174581138 0.26006831  0.15227740
#>    51   18 0.344393440 0.406971147 0.214029704 0.54672410  0.24219553
#>    18   50 0.214029704 0.546724104 0.104058003 0.51264965  0.15239185
#>    35   33 0.162013222 0.720830723 0.001873173 0.90857491  0.04546404
#>    33   34 0.001873173 0.908574909 0.041776370 0.56508597 -0.24708594
#>    34   50 0.041776370 0.565085972 0.104058003 0.51264965  0.11607785
#>    51   27 0.344393440 0.406971147 0.424206020 0.53616335  0.31767492
#>    27   31 0.424206020 0.536163348 0.379803082 0.65350974  0.32024622
#>    92   56 0.554546492 0.038762752 0.482643214 0.09860354  0.50481171
#>    91   29 0.691650581 0.615709582 0.650107487 0.73450459  0.61973464
#>    29   81 0.650107487 0.734504592 0.617592474 0.56701190  0.57895157
#>    65   71 0.973666031 0.278842853 1.000000000 0.28768139  0.96569739
#>    60   37 0.984024325 0.699756100 0.930036397 0.56546045  1.15017088
#>    93   65 0.921484029 0.405522308 0.973666031 0.27884285  0.95855615
#>    81   57 0.617592474 0.567011902 0.556635107 0.54701913  0.59261181
#>    57   38 0.556635107 0.547019127 0.540800276 0.51569295  0.57850123
#>    38   87 0.540800276 0.515692947 0.515999045 0.49427728  0.57533079
#>    87   62 0.515999045 0.494277276 0.532740528 0.36415667  0.57791569
#>    62   73 0.532740528 0.364156667 0.671161073 0.29610779  0.63755158
#>    73   82 0.671161073 0.296107786 0.749574876 0.43076895  0.64147900
#>    74   29 0.722481668 0.777350173 0.650107487 0.73450459  0.71277508
#>    85   39 0.383239436 0.795268232 0.336163379 0.72265568  0.38545898
#>    31   99 0.379803082 0.653509737 0.440038040 0.60424141  0.43421926
#>    99   36 0.440038040 0.604241409 0.574952148 0.78407847  0.54140235
#>    45   85 0.498057006 0.783784948 0.383239436 0.79526823  0.43894712
#>    94   26 0.408149833 0.950695864 0.462462275 0.85645042  0.41460293
#>    96   75 0.479997398 0.871080130 0.504816864 0.91302695  0.48396943
#>    94   85 0.408149833 0.950695864 0.383239436 0.79526823  0.38515433
#>    78   99 0.448653015 0.574447465 0.440038040 0.60424141  0.40220570
#>    51   89 0.344393440 0.406971147 0.472572830 0.40896923  0.40771729
#>    89   87 0.472572830 0.408969225 0.515999045 0.49427728  0.45111138
#>    87   58 0.515999045 0.494277276 0.456032862 0.56003129  0.46808715
#>    58   78 0.456032862 0.560031289 0.448653015 0.57444747  0.43313728
#>    64   98 0.616336186 0.947483426 0.763896003 0.98390819  0.66456936
#>    74   88 0.722481668 0.777350173 0.720985658 0.85311215  0.65200183
#>    88   36 0.720985658 0.853112153 0.574952148 0.78407847  0.64950471
#>    36   81 0.574952148 0.784078468 0.617592474 0.56701190  0.55148597
#>    83   49 0.000000000 0.016687554 0.118024454 0.14746236  0.01018317
#>    49   70 0.118024454 0.147462364 0.073946410 0.28832545  0.09717984
#>    55   34 0.067013876 0.462834518 0.041776370 0.56508597 -0.33295895
#>    33   83 0.001873173 0.908574909 0.000000000 0.01668755 -2.94738923
#>    61   98 0.849183904 0.909277531 0.763896003 0.98390819  0.78606848
#>    98   67 0.763896003 0.983908185 0.686674103 0.89816033  0.69940664
#>    67   88 0.686674103 0.898160327 0.720985658 0.85311215  0.70797473
#>    88   61 0.720985658 0.853112153 0.849183904 0.90927753  0.78906227
#>    35   43 0.162013222 0.720830723 0.213827383 0.79152778  0.09401698
#>    43   59 0.213827383 0.791527783 0.241427421 0.88024587  0.11873555
#>    59   40 0.241427421 0.880245866 0.226823618 0.93053059  0.11899438
#>    69   43 0.210993918 0.755925750 0.213827383 0.79152778  0.15807777
#>    99   57 0.440038040 0.604241409 0.556635107 0.54701913  0.54331813
#>    67   54 0.686674103 0.898160327 0.564440741 0.85962301  0.64043780
#>    57   58 0.556635107 0.547019127 0.456032862 0.56003129  0.50624354
#>    85   43 0.383239436 0.795268232 0.213827383 0.79152778  0.29841648
#>    59   94 0.241427421 0.880245866 0.408149833 0.95069586  0.31867338
#>   100   47 0.822881876 0.405485984 0.853015990 0.22867220  0.79954497
#>    47   65 0.853015990 0.228672201 0.973666031 0.27884285  0.89239433
#>    62   90 0.532740528 0.364156667 0.428306820 0.21067851  0.44718967
#>    90   66 0.428306820 0.210678506 0.496843312 0.13804491  0.51527397
#>    66   84 0.496843312 0.138044906 0.638477752 0.20243129  0.54770608
#>    84   73 0.638477752 0.202431290 0.671161073 0.29610779  0.60701179
#>    85   59 0.383239436 0.795268232 0.241427421 0.88024587  0.29807921
#>    79   61 0.938192274 0.925219504 0.849183904 0.90927753  0.88448239
#>    61   48 0.849183904 0.909277531 0.864581824 0.69655025  0.83286059
#>   100   73 0.822881876 0.405485984 0.671161073 0.29610779  0.77861882
#>    73   77 0.671161073 0.296107786 0.678954020 0.25015460  0.77217912
#>    77   68 0.678954020 0.250154597 0.794296087 0.16386211  0.76699576
#>    61   74 0.849183904 0.909277531 0.722481668 0.77735017  0.82993020
#>    83   80 0.000000000 0.016687554 0.165016429 0.14403433  0.13206589
#>    90   89 0.428306820 0.210678506 0.472572830 0.40896923  0.40987188
#>    68   76 0.794296087 0.163862114 0.691375983 0.04108218  0.76569874
#>    67   64 0.686674103 0.898160327 0.616336186 0.94748343  0.62405084
#>    92   66 0.554546492 0.038762752 0.496843312 0.13804491  0.54373154
#>    66   86 0.496843312 0.138044906 0.363112798 0.13393470  0.42985835
#>    79   98 0.938192274 0.925219504 0.763896003 0.98390819  0.87451832
#>    92   84 0.554546492 0.038762752 0.638477752 0.20243129  0.58831658
#>    77   84 0.678954020 0.250154597 0.638477752 0.20243129  0.71520709
#>    84   76 0.638477752 0.202431290 0.691375983 0.04108218  0.70439410
#>    82  100 0.749574876 0.430768950 0.822881876 0.40548598  0.79039192
#>     1   40 0.031755049 0.991370308 0.226823618 0.93053059  0.11483212
#>     1   94 0.031755049 0.991370308 0.408149833 0.95069586  0.26635253
#>     1   28 0.031755049 0.991370308 0.553432315 1.00000000  0.28330374
#>     2   13 0.288480320 0.688362462 0.246381842 0.70344020  0.28465173
#>     2   18 0.288480320 0.688362462 0.214029704 0.54672410  0.24125031
#>     2   31 0.288480320 0.688362462 0.379803082 0.65350974  0.30332009
#>     2   39 0.288480320 0.688362462 0.336163379 0.72265568  0.33519726
#>     3   72 0.672475017 0.539371877 0.826325622 0.44956761  0.76718674
#>     3   44 0.672475017 0.539371877 0.808957076 0.65902462  0.78262684
#>     3   91 0.672475017 0.539371877 0.691650581 0.61570958  0.78162959
#>     3   20 0.672475017 0.539371877 0.633743571 0.57537483  0.67378808
#>     3   81 0.672475017 0.539371877 0.617592474 0.56701190  0.64008109
#>     3   22 0.672475017 0.539371877 0.615968249 0.51206375  0.63784370
#>     3   82 0.672475017 0.539371877 0.749574876 0.43076895  0.67591211
#>     4   19 0.428193108 0.065507380 0.458748142 0.02909651  0.40886011
#>     4   56 0.428193108 0.065507380 0.482643214 0.09860354  0.46537132
#>     4   86 0.428193108 0.065507380 0.363112798 0.13393470  0.42672056
#>     4   63 0.428193108 0.065507380 0.357875639 0.02088028  0.37164855
#>     5   94 0.517615223 0.968354090 0.408149833 0.95069586  0.44179283
#>     5   25 0.517615223 0.968354090 0.467919338 0.91784026  0.46039503
#>     5   75 0.517615223 0.968354090 0.504816864 0.91302695  0.49029422
#>     5    7 0.517615223 0.968354090 0.536140570 0.95625271  0.51255701
#>     5   28 0.517615223 0.968354090 0.553432315 1.00000000  0.53876795
#>     6   46 0.993840399 0.744271671 0.880230935 0.89970752  0.94611366
#>     6   48 0.993840399 0.744271671 0.864581824 0.69655025  0.90134103
#>     6   60 0.993840399 0.744271671 0.984024325 0.69975610  0.92327405
#>     6   21 0.993840399 0.744271671 0.982453680 0.41199465  1.76163219
#>     6   71 0.993840399 0.744271671 1.000000000 0.28768139  2.29208431
#>     7   54 0.536140570 0.956252714 0.564440741 0.85962301  0.55530148
#>     7   64 0.536140570 0.956252714 0.616336186 0.94748343  0.57213421
#>     8   83 0.644438872 0.007552063 0.000000000 0.01668755  0.24381381
#>     8   53 0.644438872 0.007552063 0.784848368 0.00000000  0.18161586
#>     8   76 0.644438872 0.007552063 0.691375983 0.04108218  0.71239860
#>     8   92 0.644438872 0.007552063 0.554546492 0.03876275  0.62213974
#>     8   19 0.644438872 0.007552063 0.458748142 0.02909651  0.52839605
#>     8   63 0.644438872 0.007552063 0.357875639 0.02088028  0.46772959
#>     9   80 0.184110510 0.148647541 0.165016429 0.14403433  0.16092692
#>     9   63 0.184110510 0.148647541 0.357875639 0.02088028  0.20961021
#>     9   86 0.184110510 0.148647541 0.363112798 0.13393470  0.26871030
#>     9   10 0.184110510 0.148647541 0.223308194 0.20740115  0.27284006
#>     9   97 0.184110510 0.148647541 0.174581138 0.26006831  0.16595452
#>    10   90 0.223308194 0.207401146 0.428306820 0.21067851  0.32544043
#>    10   51 0.223308194 0.207401146 0.344393440 0.40697115  0.32463401
#>    11   96 0.503726651 0.790823745 0.479997398 0.87108013  0.51862542
#>    11   26 0.503726651 0.790823745 0.462462275 0.85645042  0.49727182
#>    11   45 0.503726651 0.790823745 0.498057006 0.78378495  0.46775747
#>    11   36 0.503726651 0.790823745 0.574952148 0.78407847  0.53660150
#>    11   54 0.503726651 0.790823745 0.564440741 0.85962301  0.54223529
#>    12  100 0.862408458 0.438332513 0.822881876 0.40548598  0.83890545
#>    12   41 0.862408458 0.438332513 0.863748525 0.38427948  0.85169106
#>    12   93 0.862408458 0.438332513 0.921484029 0.40552231  0.88636785
#>    12   37 0.862408458 0.438332513 0.930036397 0.56546045  0.92720633
#>    12   72 0.862408458 0.438332513 0.826325622 0.44956761  0.86721461
#>    13   69 0.246381842 0.703440202 0.210993918 0.75592575  0.28026893
#>    13   35 0.246381842 0.703440202 0.162013222 0.72083072  0.20444533
#>    14   65 0.909806237 0.076148862 0.973666031 0.27884285  0.96910883
#>    14   95 0.909806237 0.076148862 0.849214684 0.20794312  0.95056684
#>    14   68 0.909806237 0.076148862 0.794296087 0.16386211  0.86303655
#>    14   53 0.909806237 0.076148862 0.784848368 0.00000000  0.82173116
#>    15   74 0.835886039 0.682174470 0.722481668 0.77735017  0.81100185
#>    15   23 0.835886039 0.682174470 0.767863296 0.67330541  0.79295272
#>    15   44 0.835886039 0.682174470 0.808957076 0.65902462  0.79929864
#>    15   37 0.835886039 0.682174470 0.930036397 0.56546045  0.87096053
#>    15   48 0.835886039 0.682174470 0.864581824 0.69655025  0.88303875
#>    16   55 0.145038066 0.489797671 0.067013876 0.46283452  0.10779847
#>    16   70 0.145038066 0.489797671 0.073946410 0.28832545  0.13988497
#>    16   97 0.145038066 0.489797671 0.174581138 0.26006831  0.15227740
#>    16   51 0.145038066 0.489797671 0.344393440 0.40697115  0.21726877
#>    16   18 0.145038066 0.489797671 0.214029704 0.54672410  0.24219553
#>    16   50 0.145038066 0.489797671 0.104058003 0.51264965  0.15239185
#>    17   33 0.069881484 0.653487721 0.001873173 0.90857491  0.04546404
#>    17   34 0.069881484 0.653487721 0.041776370 0.56508597 -0.24708594
#>    17   50 0.069881484 0.653487721 0.104058003 0.51264965  0.11607785
#>    17   18 0.069881484 0.653487721 0.214029704 0.54672410  0.13862008
#>    17   35 0.069881484 0.653487721 0.162013222 0.72083072  0.16089107
#>    18   27 0.214029704 0.546724104 0.424206020 0.53616335  0.31767492
#>    20   29 0.633743571 0.575374832 0.650107487 0.73450459  0.61973464
#>    21   37 0.982453680 0.411994653 0.930036397 0.56546045  1.15017088
#>    21   93 0.982453680 0.411994653 0.921484029 0.40552231  0.94392625
#>    21   65 0.982453680 0.411994653 0.973666031 0.27884285  0.95855615
#>    22   57 0.615968249 0.512063751 0.556635107 0.54701913  0.59261181
#>    22   38 0.615968249 0.512063751 0.540800276 0.51569295  0.57850123
#>    22   87 0.615968249 0.512063751 0.515999045 0.49427728  0.57533079
#>    22   62 0.615968249 0.512063751 0.532740528 0.36415667  0.57791569
#>    22   73 0.615968249 0.512063751 0.671161073 0.29610779  0.63755158
#>    23   29 0.767863296 0.673305408 0.650107487 0.73450459  0.71277508
#>    23   91 0.767863296 0.673305408 0.691650581 0.61570958  0.69915840
#>    24   39 0.429744389 0.713051567 0.336163379 0.72265568  0.38545898
#>    24   31 0.429744389 0.713051567 0.379803082 0.65350974  0.38141633
#>    24   99 0.429744389 0.713051567 0.440038040 0.60424141  0.43421926
#>    24   36 0.429744389 0.713051567 0.574952148 0.78407847  0.54140235
#>    24   45 0.429744389 0.713051567 0.498057006 0.78378495  0.53690928
#>    24   85 0.429744389 0.713051567 0.383239436 0.79526823  0.43894712
#>    25   26 0.467919338 0.917840261 0.462462275 0.85645042  0.41460293
#>    25   96 0.467919338 0.917840261 0.479997398 0.87108013  0.45064486
#>    26   85 0.462462275 0.856450418 0.383239436 0.79526823  0.38515433
#>    27   99 0.424206020 0.536163348 0.440038040 0.60424141  0.40220570
#>    27   89 0.424206020 0.536163348 0.472572830 0.40896923  0.40771729
#>    27   87 0.424206020 0.536163348 0.515999045 0.49427728  0.45111138
#>    27   58 0.424206020 0.536163348 0.456032862 0.56003129  0.46808715
#>    27   78 0.424206020 0.536163348 0.448653015 0.57444747  0.43313728
#>    28   98 0.553432315 1.000000000 0.763896003 0.98390819  0.66456936
#>    29   88 0.650107487 0.734504592 0.720985658 0.85311215  0.65200183
#>    29   36 0.650107487 0.734504592 0.574952148 0.78407847  0.64950471
#>    30   49 0.025298039 0.235027546 0.118024454 0.14746236  0.01018317
#>    30   70 0.025298039 0.235027546 0.073946410 0.28832545  0.09717984
#>    30   55 0.025298039 0.235027546 0.067013876 0.46283452 -0.06909290
#>    30   34 0.025298039 0.235027546 0.041776370 0.56508597 -0.33295895
#>    30   33 0.025298039 0.235027546 0.001873173 0.90857491 -2.00224230
#>    30   83 0.025298039 0.235027546 0.000000000 0.01668755 -2.94738923
#>    32   98 0.725011269 0.902000073 0.763896003 0.98390819  0.78606848
#>    32   67 0.725011269 0.902000073 0.686674103 0.89816033  0.69940664
#>    32   88 0.725011269 0.902000073 0.720985658 0.85311215  0.70797473
#>    32   61 0.725011269 0.902000073 0.849183904 0.90927753  0.78906227
#>    33   43 0.001873173 0.908574909 0.213827383 0.79152778  0.09401698
#>    33   59 0.001873173 0.908574909 0.241427421 0.88024587  0.11873555
#>    36   57 0.574952148 0.784078468 0.556635107 0.54701913  0.54331813
#>    36   67 0.574952148 0.784078468 0.686674103 0.89816033  0.64293124
#>    38   58 0.540800276 0.515692947 0.456032862 0.56003129  0.50624354
#>    39   43 0.336163379 0.722655684 0.213827383 0.79152778  0.29841648
#>    41   47 0.863748525 0.384279477 0.853015990 0.22867220  0.79954497
#>    41   65 0.863748525 0.384279477 0.973666031 0.27884285  0.89239433
#>    42   90 0.548290375 0.305654004 0.428306820 0.21067851  0.44718967
#>    42   66 0.548290375 0.305654004 0.496843312 0.13804491  0.51527397
#>    42   84 0.548290375 0.305654004 0.638477752 0.20243129  0.54770608
#>    42   73 0.548290375 0.305654004 0.671161073 0.29610779  0.60701179
#>    42   62 0.548290375 0.305654004 0.532740528 0.36415667  0.61388431
#>    46   61 0.880230935 0.899707515 0.849183904 0.90927753  0.88448239
#>    47   73 0.853015990 0.228672201 0.671161073 0.29610779  0.77861882
#>    47   77 0.853015990 0.228672201 0.678954020 0.25015460  0.77217912
#>    47   68 0.853015990 0.228672201 0.794296087 0.16386211  0.76699576
#>    47   95 0.853015990 0.228672201 0.849214684 0.20794312  0.78618839
#>    49   80 0.118024454 0.147462364 0.165016429 0.14403433  0.13206589
#>    49   97 0.118024454 0.147462364 0.174581138 0.26006831  0.14577218
#>    56   66 0.482643214 0.098603542 0.496843312 0.13804491  0.54373154
#>    57   78 0.556635107 0.547019127 0.448653015 0.57444747  0.51510888
#>    62   89 0.532740528 0.364156667 0.472572830 0.40896923  0.53549084
#>    68   84 0.794296087 0.163862114 0.638477752 0.20243129  0.71520709
#>     1   33 0.031755049 0.991370308 0.001873173 0.90857491  0.11483212
#>     6   79 0.993840399 0.744271671 0.938192274 0.92521950  0.94611366
#>    14   52 0.909806237 0.076148862 0.999217041 0.27473915  0.96910883
#>    52   71 0.999217041 0.274739148 1.000000000 0.28768139  0.98727118
#>           my1          mx2           my2 bp1 bp2
#>   0.914596707   0.11899438   0.871951581   0   0
#>   1.400412047   0.31867338   0.929942808   0   0
#>   1.557275806   0.44179283   1.090262466   0   0
#>   0.743983026   0.28026893   0.764461048   0   0
#>   0.622802142   0.18804755   0.633785207   0   0
#>   0.590175858   0.32024622   0.563899841   0   0
#>   0.673701926   0.38141633   0.702871913   0   0
#>   0.524941039   0.79039192   0.430199495   0   0
#>   0.551392652   0.82520312   0.554923153   0   0
#>   0.552530161   0.76477905   0.598165067   0   0
#>   0.579619288   0.61973464   0.657221685   0   0
#>   0.543357806   0.57895157   0.661415562   0   0
#>   0.538915210   0.59261181   0.540252230   0   0
#>   0.460142957   0.64147900   0.403552727   0   0
#>   0.018257724   0.46772959  -0.704495379   0   0
#>   0.065680431   0.50481171   0.052121646   0   0
#>   0.129269021   0.42985835   0.139884626   0   0
#>   0.076890774   0.26871030   0.081659309   0   0
#>   0.974945234   0.41460293   0.891642195   0   0
#>   0.945530190   0.48396943   0.897046035   0   0
#>   0.940380327   0.55530148   0.909405393   0   0
#>   0.980505252   0.57768764   0.965121683   0   0
#>   0.828624781   0.88448239   0.968646449   0   0
#>   0.795900060   0.83286059   0.801175077   0   0
#>   0.736492119   0.92572452   0.645193473   0   0
#>   0.551626789   1.15017088   0.554964237   0   0
#>   0.533448856   0.96569739   0.346234641   0   0
#>   0.909405393   0.52335729   0.873740679   0   0
#>   0.914335235   0.62405084   0.883670270   0   0
#>   0.965121683   0.66456936   1.069187807   0   0
#>  -5.518796229   0.17969176  -0.045569860   0   0
#>  -9.906384423   0.15453735 -11.179940140   0   1
#>  -0.037963819   0.76569874   0.083307519   0   0
#>   0.088384992   0.62181365   0.107622085   0   0
#>  -0.181613645   0.50481171   0.052121646   0   0
#>   0.202782629   0.14577218   0.204031837   0   0
#>   0.001282422   0.17969176  -0.045569860   0   0
#>   0.131903540   0.32544043   0.232000590   0   0
#>   0.203212619   0.28051984   0.309207151   0   0
#>   0.232000590   0.42975065   0.143388842   0   0
#>   0.282441729   0.40987188   0.318880169   0   0
#>   0.309207151   0.21726877   0.382322207   0   0
#>   0.838865020   0.52335729   0.873740679   0   0
#>   0.832551442   0.45064486   0.888438357   0   0
#>   0.813993573   0.44182478   0.801290604   0   0
#>   0.758540742   0.53690928   0.677908302   0   0
#>   0.818029743   0.64043780   0.831693821   0   0
#>   0.426409521   0.79039192   0.430199495   0   0
#>   0.411023681   0.79954497   0.310533961   0   0
#>   0.411883378   0.91669351   0.329461817   0   0
#>   0.485414049   0.94392625   0.484519985   0   0
#>   0.517327700   0.82520312   0.554923153   0   0
#>   0.764461048   0.28127498   0.768246049   0   0
#>   0.713337650   0.15807777   0.778050968   0   0
#>   0.633785207   0.16089107   0.625671875   0   0
#>   0.168871946   0.98727118   0.281956634   0   0
#>   0.174713698   0.93443607   0.203028279   0   0
#>   0.134472217   0.78618839   0.230213983   0   0
#>   0.080076885   0.76569874   0.083307519   0   0
#>   0.767674308   0.82993020   0.800962966   0   0
#>   0.746168291   0.71277508   0.711196882   0   0
#>   0.697497179   0.76477905   0.598165067   0   0
#>   0.614136794   0.82520312   0.554923153   0   0
#>   0.623880002   0.92572452   0.645193473   0   0
#>   0.471186953   0.05130476   0.513197487   0   0
#>   0.378337154  -0.06909290   0.370035315   0   0
#>   0.373964351   0.10965561   0.222171497   0   0
#>   0.442318489   0.31767492   0.512726836   0   0
#>   0.551155505   0.13862008   0.595602433   0   0
#>   0.783587210   0.09401698   0.825001380   0   0
#>   0.705591015  -2.00224230   0.501694060   0   0
#>   0.590132214   0.05130476   0.513197487   0   0
#>   0.512726836   0.40771729   0.457100308   0   0
#>   0.563899841   0.37314757   0.583917293   0   0
#>   0.052121646   0.54373154   0.098886793   0   0
#>   0.657221685   0.69915840   0.684996496   0   0
#>   0.661415562   0.55148597   0.666747403   0   0
#>   0.346234641   0.98727118   0.281956634   0   0
#>   0.554964237   0.92572452   0.645193473   0   0
#>   0.346705943   0.91669351   0.329461817   0   0
#>   0.540252230   0.55114588   0.666680596   0   0
#>   0.516300984   0.50624354   0.552825970   0   0
#>   0.450634738   0.49901790   0.539011769   0   0
#>   0.436106255   0.53549084   0.430647821   0   0
#>   0.402548978   0.61388431   0.354406575   0   0
#>   0.403552727   0.75783390   0.335798748   0   0
#>   0.711196882   0.65200183   0.813854226   0   0
#>   0.742262850   0.29841648   0.798694111   0   0
#>   0.658582920   0.37314757   0.583917293   0   0
#>   0.668722647   0.54331813   0.667285428   0   0
#>   0.772517830   0.44182478   0.801290604   0   0
#>   0.891642195   0.38515433   0.874671344   0   0
#>   0.897046035   0.52335729   0.873740679   0   0
#>   0.874671344   0.33889604   0.882085162   0   0
#>   0.577159626   0.51510888   0.609805794   0   0
#>   0.457100308   0.40987188   0.318880169   0   0
#>   0.473601336   0.53549084   0.430647821   0   0
#>   0.510803616   0.49901790   0.539011769   0   0
#>   0.557407725   0.51206195   0.597810391   0   0
#>   1.069187807   0.68789089   0.974710296   0   0
#>   0.813854226   0.81319313   0.817037139   0   0
#>   0.815346470   0.64293124   0.829251961   0   0
#>   0.666747403   0.55114588   0.666680596   0   0
#>   0.126143257   0.13206589   0.016143925   0   0
#>   0.218267652   0.10965561   0.222171497   0   0
#>   0.418354261   0.05130476   0.513197487   0   0
#>   0.468823410  -4.06860211   0.471178221   0   1
#>   0.923198062   0.87451832   1.024278531   0   0
#>   0.964339561   0.68789089   0.974710296   0   0
#>   0.878793223   0.64293124   0.829251961   0   0
#>   0.872116176   0.81319313   0.817037139   0   0
#>   0.825001380   0.15807777   0.778050968   0   0
#>   0.869762886   0.29807921   0.813969392   0   0
#>   0.871951581   0.31867338   0.929942808   0   0
#>   0.778050968   0.28127498   0.768246049   0   0
#>   0.667285428   0.51510888   0.609805794   0   0
#>   0.831693821   0.62405084   0.883670270   0   0
#>   0.552825970   0.51206195   0.597810391   0   0
#>   0.798694111   0.29807921   0.813969392   0   0
#>   0.929942808   0.33889604   0.882085162   0   0
#>   0.310533961   0.77861882   0.306967548   0   0
#>   0.304129960   0.93443607   0.203028279   0   0
#>   0.310099599   0.44620582   0.310769056   0   0
#>   0.224087978   0.42975065   0.143388842   0   0
#>   0.214133047   0.58831658   0.124799796   0   0
#>   0.265949411   0.61541659   0.263017012   0   0
#>   0.813969392   0.33889604   0.882085162   0   0
#>   0.968646449   0.87451832   1.024278531   0   0
#>   0.801175077   0.82993020   0.800962966   0   0
#>   0.306967548   0.75783390   0.335798748   0   0
#>   0.289601502   0.61541659   0.263017012   0   0
#>   0.247603086   0.71520709   0.178380229   0   0
#>   0.800962966   0.81319313   0.817037139   0   0
#>   0.016143925   0.17969176  -0.045569860   0   0
#>   0.318880169   0.44620582   0.310769056   0   0
#>   0.083307519   0.70439410   0.134696047   0   0
#>   0.883670270   0.68789089   0.974710296   0   0
#>   0.098886793   0.58831658   0.124799796   0   0
#>   0.139884626   0.42975065   0.143388842   0   0
#>   1.024278531   2.60966083   6.177382314   0   1
#>   0.124799796   0.62181365   0.107622085   0   0
#>   0.178380229   0.61541659   0.263017012   0   0
#>   0.134696047   0.62181365   0.107622085   0   0
#>   0.430199495   0.75783390   0.335798748   0   0
#>   0.914596707   0.26635253   1.400412047   0   0
#>   1.400412047   0.28330374   1.557275806   0   0
#>   1.557275806   0.25160276   3.473645363   0   1
#>   0.743983026   0.24125031   0.622802142   0   0
#>   0.622802142   0.30332009   0.590175858   0   0
#>   0.590175858   0.33519726   0.673701926   0   0
#>   0.673701926   0.28465173   0.743983026   0   0
#>   0.524941039   0.78262684   0.551392652   0   0
#>   0.551392652   0.78162959   0.552530161   0   0
#>   0.552530161   0.67378808   0.579619288   0   0
#>   0.579619288   0.64008109   0.543357806   0   0
#>   0.543357806   0.63784370   0.538915210   0   0
#>   0.538915210   0.67591211   0.460142957   0   0
#>   0.460142957   0.76718674   0.524941039   0   0
#>   0.018257724   0.46537132   0.065680431   0   0
#>   0.065680431   0.42672056   0.129269021   0   0
#>   0.129269021   0.37164855   0.076890774   0   0
#>   0.076890774   0.40886011   0.018257724   0   0
#>   1.090262466   0.46039503   0.974945234   0   0
#>   0.974945234   0.49029422   0.945530190   0   0
#>   0.945530190   0.51255701   0.940380327   0   0
#>   0.940380327   0.53876795   0.980505252   0   0
#>   0.980505252   0.44179283   1.090262466   0   0
#>   0.828624781   0.90134103   0.795900060   0   0
#>   0.795900060   0.92327405   0.736492119   0   0
#>   0.736492119   1.76163219   0.551626789   0   0
#>   0.551626789   2.29208431   0.533448856   0   0
#>   0.533448856   4.48183316   0.562989519   0   1
#>   0.909405393   0.57213421   0.914335235   0   0
#>   0.914335235   0.57768764   0.965121683   0   0
#>  -5.518796229   0.18161586  -9.906384423   0   0
#>  -9.906384423   0.71239860  -0.037963819   0   0
#>  -0.037963819   0.62213974   0.088384992   0   0
#>   0.088384992   0.52839605  -0.181613645   0   0
#>  -0.181613645   0.46772959  -0.704495379   0   0
#>  -0.704495379   0.24381381  -5.518796229   0   0
#>   0.202782629   0.20961021   0.001282422   0   0
#>   0.001282422   0.26871030   0.081659309   0   0
#>   0.081659309   0.27284006   0.131903540   0   0
#>   0.131903540   0.16595452   0.203212619   0   0
#>   0.203212619   0.16092692   0.202782629   0   0
#>   0.232000590   0.32463401   0.282441729   0   0
#>   0.282441729   0.28051984   0.309207151   0   0
#>   0.838865020   0.49727182   0.832551442   0   0
#>   0.832551442   0.46775747   0.813993573   0   0
#>   0.813993573   0.53660150   0.758540742   0   0
#>   0.758540742   0.54223529   0.818029743   0   0
#>   0.818029743   0.51862542   0.838865020   0   0
#>   0.426409521   0.85169106   0.411023681   0   0
#>   0.411023681   0.88636785   0.411883378   0   0
#>   0.411883378   0.92720633   0.485414049   0   0
#>   0.485414049   0.86721461   0.517327700   0   0
#>   0.517327700   0.83890545   0.426409521   0   0
#>   0.764461048   0.20444533   0.713337650   0   0
#>   0.713337650   0.18804755   0.633785207   0   0
#>   0.168871946   0.95056684   0.174713698   0   0
#>   0.174713698   0.86303655   0.134472217   0   0
#>   0.134472217   0.82173116   0.080076885   0   0
#>   0.080076885   4.37790927  -5.755498679   0   1
#>   0.767674308   0.79295272   0.746168291   0   0
#>   0.746168291   0.79929864   0.697497179   0   0
#>   0.697497179   0.87096053   0.614136794   0   0
#>   0.614136794   0.88303875   0.623880002   0   0
#>   0.623880002   0.81100185   0.767674308   0   0
#>   0.471186953   0.13988497   0.378337154   0   0
#>   0.378337154   0.15227740   0.373964351   0   0
#>   0.373964351   0.21726877   0.382322207   0   0
#>   0.382322207   0.24219553   0.442318489   0   0
#>   0.442318489   0.15239185   0.551155505   0   0
#>   0.551155505   0.10779847   0.471186953   0   0
#>   0.783587210  -0.24708594   0.705591015   0   0
#>   0.705591015   0.11607785   0.590132214   0   0
#>   0.590132214   0.13862008   0.595602433   0   0
#>   0.595602433   0.16089107   0.625671875   0   0
#>   0.625671875   0.04546404   0.783587210   0   0
#>   0.512726836   0.32024622   0.563899841   0   0
#>   0.657221685   0.57895157   0.661415562   0   0
#>   0.554964237   0.94392625   0.484519985   0   0
#>   0.484519985   0.95855615   0.346705943   0   0
#>   0.346705943   0.96569739   0.346234641   0   0
#>   0.540252230   0.57850123   0.516300984   0   0
#>   0.516300984   0.57533079   0.450634738   0   0
#>   0.450634738   0.57791569   0.436106255   0   0
#>   0.436106255   0.63755158   0.402548978   0   0
#>   0.402548978   0.64147900   0.403552727   0   0
#>   0.711196882   0.69915840   0.684996496   0   0
#>   0.684996496   0.76477905   0.598165067   0   0
#>   0.742262850   0.38141633   0.702871913   0   0
#>   0.702871913   0.43421926   0.658582920   0   0
#>   0.658582920   0.54140235   0.668722647   0   0
#>   0.668722647   0.53690928   0.677908302   0   0
#>   0.677908302   0.43894712   0.772517830   0   0
#>   0.772517830   0.38545898   0.742262850   0   0
#>   0.891642195   0.45064486   0.888438357   0   0
#>   0.888438357   0.48396943   0.897046035   0   0
#>   0.874671344   0.44182478   0.801290604   0   0
#>   0.577159626   0.37314757   0.583917293   0   0
#>   0.457100308   0.45111138   0.473601336   0   0
#>   0.473601336   0.46808715   0.510803616   0   0
#>   0.510803616   0.43313728   0.557407725   0   0
#>   0.557407725   0.40220570   0.577159626   0   0
#>   1.069187807   1.02574562   5.792986138   0   1
#>   0.813854226   0.64950471   0.815346470   0   0
#>   0.815346470   0.55148597   0.666747403   0   0
#>   0.126143257   0.09717984   0.218267652   0   0
#>   0.218267652  -0.06909290   0.370035315   0   0
#>   0.370035315  -0.33295895   0.418354261   0   0
#>   0.418354261  -2.00224230   0.501694060   0   0
#>   0.501694060  -2.94738923   0.468823410   0   0
#>   0.468823410   0.01018317   0.126143257   0   0
#>   0.923198062   0.69940664   0.964339561   0   0
#>   0.964339561   0.70797473   0.878793223   0   0
#>   0.878793223   0.78906227   0.872116176   0   0
#>   0.872116176   0.78606848   0.923198062   0   0
#>   0.825001380   0.11873555   0.869762886   0   0
#>   0.869762886   0.11899438   0.871951581   0   0
#>   0.667285428   0.55114588   0.666680596   0   0
#>   0.829251961   0.64043780   0.831693821   0   0
#>   0.552825970   0.49901790   0.539011769   0   0
#>   0.798694111   0.28127498   0.768246049   0   0
#>   0.310533961   0.89239433   0.304129960   0   0
#>   0.304129960   0.91669351   0.329461817   0   0
#>   0.310099599   0.51527397   0.224087978   0   0
#>   0.224087978   0.54770608   0.214133047   0   0
#>   0.214133047   0.60701179   0.265949411   0   0
#>   0.265949411   0.61388431   0.354406575   0   0
#>   0.354406575   0.44718967   0.310099599   0   0
#>   0.968646449   0.83286059   0.801175077   0   0
#>   0.306967548   0.77217912   0.289601502   0   0
#>   0.289601502   0.76699576   0.247603086   0   0
#>   0.247603086   0.78618839   0.230213983   0   0
#>   0.230213983   0.93443607   0.203028279   0   0
#>   0.016143925   0.14577218   0.204031837   0   0
#>   0.204031837   0.10965561   0.222171497   0   0
#>   0.098886793   0.42985835   0.139884626   0   0
#>   0.609805794   0.51206195   0.597810391   0   0
#>   0.430647821   0.44620582   0.310769056   0   0
#>   0.178380229   0.70439410   0.134696047   0   0
#>   0.914596707 -10.57119643   4.771315567   0   1
#>   0.828624781   5.99504587   2.381357433   0   1
#>   0.168871946   5.15591602  -1.716143699   0   1
#>   0.281956634  77.97188758  -4.375337599   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.031755049 0.991370308
#>   [2,] 0.288480320 0.688362462
#>   [3,] 0.672475017 0.539371877
#>   [4,] 0.428193108 0.065507380
#>   [5,] 0.517615223 0.968354090
#>   [6,] 0.993840399 0.744271671
#>   [7,] 0.536140570 0.956252714
#>   [8,] 0.644438872 0.007552063
#>   [9,] 0.184110510 0.148647541
#>  [10,] 0.223308194 0.207401146
#>  [11,] 0.503726651 0.790823745
#>  [12,] 0.862408458 0.438332513
#>  [13,] 0.246381842 0.703440202
#>  [14,] 0.909806237 0.076148862
#>  [15,] 0.835886039 0.682174470
#>  [16,] 0.145038066 0.489797671
#>  [17,] 0.069881484 0.653487721
#>  [18,] 0.214029704 0.546724104
#>  [19,] 0.458748142 0.029096514
#>  [20,] 0.633743571 0.575374832
#>  [21,] 0.982453680 0.411994653
#>  [22,] 0.615968249 0.512063751
#>  [23,] 0.767863296 0.673305408
#>  [24,] 0.429744389 0.713051567
#>  [25,] 0.467919338 0.917840261
#>  [26,] 0.462462275 0.856450418
#>  [27,] 0.424206020 0.536163348
#>  [28,] 0.553432315 1.000000000
#>  [29,] 0.650107487 0.734504592
#>  [30,] 0.025298039 0.235027546
#>  [31,] 0.379803082 0.653509737
#>  [32,] 0.725011269 0.902000073
#>  [33,] 0.001873173 0.908574909
#>  [34,] 0.041776370 0.565085972
#>  [35,] 0.162013222 0.720830723
#>  [36,] 0.574952148 0.784078468
#>  [37,] 0.930036397 0.565460447
#>  [38,] 0.540800276 0.515692947
#>  [39,] 0.336163379 0.722655684
#>  [40,] 0.226823618 0.930530587
#>  [41,] 0.863748525 0.384279477
#>  [42,] 0.548290375 0.305654004
#>  [43,] 0.213827383 0.791527783
#>  [44,] 0.808957076 0.659024623
#>  [45,] 0.498057006 0.783784948
#>  [46,] 0.880230935 0.899707515
#>  [47,] 0.853015990 0.228672201
#>  [48,] 0.864581824 0.696550251
#>  [49,] 0.118024454 0.147462364
#>  [50,] 0.104058003 0.512649649
#>  [51,] 0.344393440 0.406971147
#>  [52,] 0.999217041 0.274739148
#>  [53,] 0.784848368 0.000000000
#>  [54,] 0.564440741 0.859623008
#>  [55,] 0.067013876 0.462834518
#>  [56,] 0.482643214 0.098603542
#>  [57,] 0.556635107 0.547019127
#>  [58,] 0.456032862 0.560031289
#>  [59,] 0.241427421 0.880245866
#>  [60,] 0.984024325 0.699756100
#>  [61,] 0.849183904 0.909277531
#>  [62,] 0.532740528 0.364156667
#>  [63,] 0.357875639 0.020880277
#>  [64,] 0.616336186 0.947483426
#>  [65,] 0.973666031 0.278842853
#>  [66,] 0.496843312 0.138044906
#>  [67,] 0.686674103 0.898160327
#>  [68,] 0.794296087 0.163862114
#>  [69,] 0.210993918 0.755925750
#>  [70,] 0.073946410 0.288325448
#>  [71,] 1.000000000 0.287681390
#>  [72,] 0.826325622 0.449567612
#>  [73,] 0.671161073 0.296107786
#>  [74,] 0.722481668 0.777350173
#>  [75,] 0.504816864 0.913026953
#>  [76,] 0.691375983 0.041082179
#>  [77,] 0.678954020 0.250154597
#>  [78,] 0.448653015 0.574447465
#>  [79,] 0.938192274 0.925219504
#>  [80,] 0.165016429 0.144034332
#>  [81,] 0.617592474 0.567011902
#>  [82,] 0.749574876 0.430768950
#>  [83,] 0.000000000 0.016687554
#>  [84,] 0.638477752 0.202431290
#>  [85,] 0.383239436 0.795268232
#>  [86,] 0.363112798 0.133934702
#>  [87,] 0.515999045 0.494277276
#>  [88,] 0.720985658 0.853112153
#>  [89,] 0.472572830 0.408969225
#>  [90,] 0.428306820 0.210678506
#>  [91,] 0.691650581 0.615709582
#>  [92,] 0.554546492 0.038762752
#>  [93,] 0.921484029 0.405522308
#>  [94,] 0.408149833 0.950695864
#>  [95,] 0.849214684 0.207943119
#>  [96,] 0.479997398 0.871080130
#>  [97,] 0.174581138 0.260068306
#>  [98,] 0.763896003 0.983908185
#>  [99,] 0.440038040 0.604241409
#> [100,] 0.822881876 0.405485984
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.03175505,0.9913703) [4]: 28 33 40 94 
#> 2: (0.2884803,0.6883625) [4]: 13 18 31 39 
#> 3: (0.672475,0.5393719) [7]: 20 22 44 72 81 82 91 
#> 4: (0.4281931,0.06550738) [4]: 19 56 63 86 
#> 5: (0.5176152,0.9683541) [5]: 7 25 28 75 94 
#> 6: (0.9938404,0.7442717) [6]: 21 46 48 60 71 79 
#> 7: (0.5361406,0.9562527) [5]: 5 28 54 64 75 
#> 8: (0.6444389,0.007552063) [6]: 19 53 63 76 83 92 
#> 9: (0.1841105,0.1486475) [5]: 10 63 80 86 97 
#> 10: (0.2233082,0.2074011) [5]: 9 51 86 90 97 
#> 11: (0.5037267,0.7908237) [5]: 26 36 45 54 96 
#> 12: (0.8624085,0.4383325) [5]: 37 41 72 93 100 
#> 13: (0.2463818,0.7034402) [5]: 2 18 35 39 69 
#> 14: (0.9098062,0.07614886) [5]: 52 53 65 68 95 
#> 15: (0.835886,0.6821745) [5]: 23 37 44 48 74 
#> 16: (0.1450381,0.4897977) [6]: 18 50 51 55 70 97 
#> 17: (0.06988148,0.6534877) [5]: 18 33 34 35 50 
#> 18: (0.2140297,0.5467241) [9]: 2 13 16 17 27 31 35 50 51 
#> 19: (0.4587481,0.02909651) [5]: 4 8 56 63 92 
#> 20: (0.6337436,0.5753748) [4]: 3 29 81 91 
#> 21: (0.9824537,0.4119947) [6]: 6 37 60 65 71 93 
#> 22: (0.6159682,0.5120638) [8]: 3 38 57 62 73 81 82 87 
#> 23: (0.7678633,0.6733054) [5]: 15 29 44 74 91 
#> 24: (0.4297444,0.7130516) [6]: 31 36 39 45 85 99 
#> 25: (0.4679193,0.9178403) [5]: 5 26 75 94 96 
#> 26: (0.4624623,0.8564504) [6]: 11 25 45 85 94 96 
#> 27: (0.424206,0.5361633) [8]: 18 31 51 58 78 87 89 99 
#> 28: (0.5534323,1) [6]: 1 5 7 64 94 98 
#> 29: (0.6501075,0.7345046) [7]: 20 23 36 74 81 88 91 
#> 30: (0.02529804,0.2350275) [6]: 33 34 49 55 70 83 
#> 31: (0.3798031,0.6535097) [6]: 2 18 24 27 39 99 
#> 32: (0.7250113,0.9020001) [4]: 61 67 88 98 
#> 33: (0.001873173,0.9085749) [9]: 1 17 30 34 35 40 43 59 83 
#> 34: (0.04177637,0.565086) [5]: 17 30 33 50 55 
#> 35: (0.1620132,0.7208307) [6]: 13 17 18 33 43 69 
#> 36: (0.5749521,0.7840785) [10]: 11 24 29 45 54 57 67 81 88 99 
#> 37: (0.9300364,0.5654604) [8]: 12 15 21 44 48 60 72 93 
#> 38: (0.5408003,0.5156929) [4]: 22 57 58 87 
#> 39: (0.3361634,0.7226557) [7]: 2 13 24 31 43 69 85 
#> 40: (0.2268236,0.9305306) [4]: 1 33 59 94 
#> 41: (0.8637485,0.3842795) [5]: 12 47 65 93 100 
#> 42: (0.5482904,0.305654) [5]: 62 66 73 84 90 
#> 43: (0.2138274,0.7915278) [6]: 33 35 39 59 69 85 
#> 44: (0.8089571,0.6590246) [6]: 3 15 23 37 72 91 
#> 45: (0.498057,0.7837849) [5]: 11 24 26 36 85 
#> 46: (0.8802309,0.8997075) [4]: 6 48 61 79 
#> 47: (0.853016,0.2286722) [7]: 41 65 68 73 77 95 100 
#> 48: (0.8645818,0.6965503) [7]: 6 15 37 46 60 61 74 
#> 49: (0.1180245,0.1474624) [5]: 30 70 80 83 97 
#> 50: (0.104058,0.5126496) [5]: 16 17 18 34 55 
#> 51: (0.3443934,0.4069711) [7]: 10 16 18 27 89 90 97 
#> 52: (0.999217,0.2747391) [3]: 14 65 71 
#> 53: (0.7848484,0) [5]: 8 14 68 76 83 
#> 54: (0.5644407,0.859623) [7]: 7 11 36 64 67 75 96 
#> 55: (0.06701388,0.4628345) [5]: 16 30 34 50 70 
#> 56: (0.4826432,0.09860354) [5]: 4 19 66 86 92 
#> 57: (0.5566351,0.5470191) [7]: 22 36 38 58 78 81 99 
#> 58: (0.4560329,0.5600313) [5]: 27 38 57 78 87 
#> 59: (0.2414274,0.8802459) [5]: 33 40 43 85 94 
#> 60: (0.9840243,0.6997561) [4]: 6 21 37 48 
#> 61: (0.8491839,0.9092775) [7]: 32 46 48 74 79 88 98 
#> 62: (0.5327405,0.3641567) [6]: 22 42 73 87 89 90 
#> 63: (0.3578756,0.02088028) [7]: 4 8 9 19 80 83 86 
#> 64: (0.6163362,0.9474834) [5]: 7 28 54 67 98 
#> 65: (0.973666,0.2788429) [8]: 14 21 41 47 52 71 93 95 
#> 66: (0.4968433,0.1380449) [6]: 42 56 84 86 90 92 
#> 67: (0.6866741,0.8981603) [6]: 32 36 54 64 88 98 
#> 68: (0.7942961,0.1638621) [7]: 14 47 53 76 77 84 95 
#> 69: (0.2109939,0.7559258) [4]: 13 35 39 43 
#> 70: (0.07394641,0.2883254) [5]: 16 30 49 55 97 
#> 71: (1,0.2876814) [4]: 6 21 52 65 
#> 72: (0.8263256,0.4495676) [6]: 3 12 37 44 82 100 
#> 73: (0.6711611,0.2961078) [8]: 22 42 47 62 77 82 84 100 
#> 74: (0.7224817,0.7773502) [6]: 15 23 29 48 61 88 
#> 75: (0.5048169,0.913027) [5]: 5 7 25 54 96 
#> 76: (0.691376,0.04108218) [5]: 8 53 68 84 92 
#> 77: (0.678954,0.2501546) [4]: 47 68 73 84 
#> 78: (0.448653,0.5744475) [4]: 27 57 58 99 
#> 79: (0.9381923,0.9252195) [4]: 6 46 61 98 
#> 80: (0.1650164,0.1440343) [5]: 9 49 63 83 97 
#> 81: (0.6175925,0.5670119) [6]: 3 20 22 29 36 57 
#> 82: (0.7495749,0.4307689) [5]: 3 22 72 73 100 
#> 83: (0,0.01668755) [7]: 8 30 33 49 53 63 80 
#> 84: (0.6384778,0.2024313) [7]: 42 66 68 73 76 77 92 
#> 85: (0.3832394,0.7952682) [7]: 24 26 39 43 45 59 94 
#> 86: (0.3631128,0.1339347) [7]: 4 9 10 56 63 66 90 
#> 87: (0.515999,0.4942773) [6]: 22 27 38 58 62 89 
#> 88: (0.7209857,0.8531122) [6]: 29 32 36 61 67 74 
#> 89: (0.4725728,0.4089692) [5]: 27 51 62 87 90 
#> 90: (0.4283068,0.2106785) [7]: 10 42 51 62 66 86 89 
#> 91: (0.6916506,0.6157096) [5]: 3 20 23 29 44 
#> 92: (0.5545465,0.03876275) [6]: 8 19 56 66 76 84 
#> 93: (0.921484,0.4055223) [5]: 12 21 37 41 65 
#> 94: (0.4081498,0.9506959) [8]: 1 5 25 26 28 40 59 85 
#> 95: (0.8492147,0.2079431) [4]: 14 47 65 68 
#> 96: (0.4799974,0.8710801) [5]: 11 25 26 54 75 
#> 97: (0.1745811,0.2600683) [7]: 9 10 16 49 51 70 80 
#> 98: (0.763896,0.9839082) [6]: 28 32 61 64 67 79 
#> 99: (0.440038,0.6042414) [6]: 24 27 31 36 57 78 
#> 100: (0.8228819,0.405486) [6]: 12 41 47 72 73 82 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  1 6 14 28 33 52 53 71 79 83 98 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.226019368 0.182444058 0.153420655 0.091814809 0.160020612 0.197190223 
#>                                                                         
#> 0.081765430 0.079019407 0.210175892 0.125048017 0.070569921 0.018187813 
#>                                                                         
#> 0.054972151 0.156395573 0.101206563 0.073499670 0.124642734 0.113175664 
#>                                                                         
#> 0.068204725 0.037210151 0.053382029 0.047040721 0.063327548 0.203759095 
#>                                                                         
#> 0.119485516 0.287765733 0.125545453 0.080043663 0.102042100 0.081944417 
#>                                                                         
#> 0.357900198 0.785025755 0.102102068 0.136849149 0.096284786 0.116427518 
#>                                                                         
#> 0.228826563 0.157932409 0.071750650 0.100696931 0.213476594 0.224536551 
#>                                                                         
#> 0.085217040 0.022836572 0.080915112 0.076895702 0.076272323 0.044215939 
#>                                                                         
#> 0.046041274 0.061519479 0.160166636 0.155521941 0.129515603 0.060255867 
#>                                                                         
#> 0.181710839 0.025878456 0.143230269 0.070421498 0.164134249 0.163465843 
#>                                                                         
#> 0.113511257 0.043504478 0.153017832 0.146522482 0.062079100 0.174646717 
#>                                                                         
#> 0.104526621 0.191116699 0.115129681 0.246764492 0.345798951 0.081416030 
#>                                                                         
#> 0.151857410 0.125466314 0.093546788 0.125849446 0.170619539 0.027777647 
#>                                                                         
#> 0.144741213 0.137006006 0.064152254 0.035100875 0.032767850 0.131193178 
#>                                                                         
#> 0.154242982 0.155827961 0.084105683 0.086537491 0.077817854 0.224818114 
#>                                                                         
#> 0.115390381 0.108775207 0.048739531 0.157411171 0.031014462 0.128194963 
#>                                                                         
#> 0.095725125 0.088991760 0.016195317 0.151989022 0.075776749 0.161528436 
#>                                                                         
#> 0.221215034 0.176158515 0.147598382 0.105319949 0.891889321 0.113330316 
#>                                                                         
#> 0.115394614 0.056627032 0.139961924 0.087651478 0.092912111 0.052362432 
#>                                                                         
#> 0.035714609 0.129881737 0.128164424 0.101440268 0.169453341 0.180996036 
#>                                                                         
#> 0.179363259 0.130665706 0.185639288 0.099864361 0.155582522 0.099214341 
#>                                                                         
#> 0.165323459 0.090424756 0.213283829 0.187036874 0.046609287 0.144049245 
#>                                                                         
#> 0.182916059 0.208440935 0.203171575 0.160210674 0.085908036 0.114832935 
#>                                                                         
#> 0.133793662 0.183911804 0.183934355 0.062576691 0.169799171 0.077544469 
#>                                                                         
#> 0.204336043 0.378586111 0.521748638 0.044717112 0.160013496 0.097747425 
#>                                                                         
#> 0.058734139 0.178142680 0.181505185 0.078709259 0.052880409 0.061449691 
#>                                                                         
#> 0.062759450 0.133187777 0.047532738 0.063719464 0.094433814 0.083283400 
#>                                                                         
#> 0.110880496 0.070861328 0.056788115 0.022127625 0.047794641 0.192528990 
#>                                                                         
#> 0.137786477 0.045584991 0.332472065 0.456631827 0.100688628 0.080673646 
#>                                                                         
#> 0.644503621 0.140612447 0.057683282 0.095156435 0.186936382 0.286873017 
#>                                                                         
#> 0.019643463 0.215682159 0.179605921 0.070628922 0.111827527 0.205024822 
#>                                                                         
#> 0.233430550 0.083690889 0.077521668 0.009038227 0.071544183 0.091758048 
#>                                                                         
#> 0.051393045 0.054069645 0.067575385 0.143996700 0.037791514 0.063301169 
#>                                                                         
#> 0.086142291 0.212515711 0.145055377 0.145038647 0.146332218 0.148050552 
#>                                                                         
#> 0.068598497 0.035511751 0.149954837 0.032095345 0.082551716 0.213647093 
#>                                                                         
#> 0.231621187 0.215876813 0.089445319 0.046920980 0.263997356 0.092761882 
#>                                                                         
#> 0.144925488 0.179379986 0.114119836 0.210441472 0.159968929 0.162170656 
#>                                                                         
#> 0.061312230 0.133441464 0.068864360 0.075255534 0.101539158 0.169715524 
#>                                                                         
#> 0.222897345 0.132709347 0.095528305 0.094072549 0.077713343 0.109295973 
#>                                                                         
#> 0.161648117 0.098335267 0.094457878 0.061631911 0.048294817 0.100097543 
#>                                                                         
#> 0.069894744 0.136079731 0.100897981 0.039782238 0.045423884 0.211077973 
#>                                                                         
#> 0.138171881 0.090032740 0.127537638 0.072161835 0.231594965 0.330469515 
#>                                                                         
#> 0.673954579 0.219800688 0.090669517 0.038528976 0.049053382 0.124385709 
#>                                                                         
#> 0.242125209 0.241223490 0.237765946 0.159676128 0.095662966 0.140390391 
#>                                                                         
#> 0.155976959 0.152311325 0.153024177 0.175327151 0.137071848 0.123240978 
#>                                                                         
#> 0.060533952 0.032488511 0.193955585 0.175382617 0.087454985 0.021074743 
#>                                                                         
#> 0.047116846 0.126010939 0.041919732 0.111411157 0.075022112 0.160520824 
#>                                                 
#> 0.088022751 0.189311468 0.217789792 0.012965903 
#> 
#> $alpha
#> [1] 0.3011734
#> 
#> attr(,"class")
#> [1] "scree"