Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1           y1           x2           y2         mx1
#>    58   39 0.1681300526 0.7800724814 0.1256119587 0.8588129805  0.10573396
#>    72    3 0.1070682540 0.6871840071 0.1837334308 0.7291765178  0.12800885
#>     1   72 0.0445426497 0.8182835048 0.1070682540 0.6871840071  0.09477365
#>    69   61 0.1999095636 0.7364870890 0.2709264618 0.7653392217  0.22136312
#>    58    3 0.1681300526 0.7800724814 0.1837334308 0.7291765178  0.18123748
#>    63   89 0.0658449161 0.6304780067 0.1797244246 0.6284064139  0.12283816
#>    58   34 0.1681300526 0.7800724814 0.2794968263 0.8003651745  0.21333384
#>    89    3 0.1797244246 0.6284064139 0.1837334308 0.7291765178  0.16104756
#>    69    3 0.1999095636 0.7364870890 0.1837334308 0.7291765178  0.21687619
#>    42   63 0.0513959401 0.6287016328 0.0658449161 0.6304780067  0.05194065
#>    61   34 0.2709264618 0.7653392217 0.2794968263 0.8003651745  0.22281984
#>     1   42 0.0445426497 0.8182835048 0.0513959401 0.6287016328  0.01174915
#>    84   48 0.2330876722 0.9266669517 0.2697440470 0.9822331923  0.06114313
#>    63  100 0.0658449161 0.6304780067 0.0000000000 0.5620638226  0.09528925
#>    84   34 0.2330876722 0.9266669517 0.2794968263 0.8003651745  0.23483071
#>    48    1 0.2697440470 0.9822331923 0.0445426497 0.8182835048 -0.17136379
#>    84   38 0.2330876722 0.9266669517 0.2592523979 0.9223598056  0.25112924
#>    99   48 0.3688672651 0.9306484457 0.2697440470 0.9822331923  0.31274885
#>    42  100 0.0513959401 0.6287016328 0.0000000000 0.5620638226  0.06671519
#>    34   99 0.2794968263 0.8003651745 0.3688672651 0.9306484457  0.31837077
#>    62   61 0.3828943585 0.6272658180 0.2709264618 0.7653392217  0.28148697
#>    89   52 0.1797244246 0.6284064139 0.1915779281 0.5144500557  0.24435519
#>    30   11 0.5135064289 0.8197022944 0.4875081331 0.8615031451  0.39918311
#>    62   89 0.3828943585 0.6272658180 0.1797244246 0.6284064139  0.28119983
#>    85   52 0.2829283602 0.5048901793 0.1915779281 0.5144500557  0.23467593
#>    74   25 0.0105203711 0.3730006507 0.2037826419 0.4429980902  0.08830252
#>    85   56 0.2829283602 0.5048901793 0.3160067188 0.4538210814  0.34883759
#>    62   85 0.3828943585 0.6272658180 0.2829283602 0.5048901793  0.34789756
#>    54   56 0.2806953197 0.4237829532 0.3160067188 0.4538210814  0.27650542
#>    62   30 0.3828943585 0.6272658180 0.5135064289 0.8197022944  0.40100666
#>    99   11 0.3688672651 0.9306484457 0.4875081331 0.8615031451  0.44019047
#>    64   44 0.5722805662 0.8877758082 0.4733020205 0.9813697625  0.51415151
#>    25   54 0.2037826419 0.4429980902 0.2806953197 0.4237829532  0.25018777
#>    95   62 0.4613836455 0.5898542784 0.3828943585 0.6272658180  0.40312255
#>    43   96 0.3485513655 0.4359982681 0.4149672393 0.5078104701  0.35872772
#>    70   94 0.0078953072 0.3644665449 0.1019776048 0.2568421801  0.09217997
#>   100   74 0.0000000000 0.5620638226 0.0105203711 0.3730006507  0.07923446
#>    54   43 0.2806953197 0.4237829532 0.3485513655 0.4359982681  0.31695642
#>    25   45 0.2037826419 0.4429980902 0.1970939521 0.2853621123  0.18431710
#>    74   94 0.0105203711 0.3730006507 0.1019776048 0.2568421801  0.12108934
#>    30   64 0.5135064289 0.8197022944 0.5722805662 0.8877758082  0.53398916
#>    30   86 0.5135064289 0.8197022944 0.5812325449 0.7744185289  0.56282594
#>    95   86 0.4613836455 0.5898542784 0.5812325449 0.7744185289  0.48590094
#>    81   86 0.6780784277 0.7572867137 0.5812325449 0.7744185289  0.61427973
#>    43   31 0.3485513655 0.4359982681 0.3018658068 0.3365209145  0.34708187
#>    31   54 0.3018658068 0.3365209145 0.2806953197 0.4237829532  0.26585401
#>    15   95 0.5633155366 0.4932438849 0.4613836455 0.5898542784  0.49095267
#>     5   54 0.2833973812 0.3250699966 0.2806953197 0.4237829532  0.23344664
#>    70   37 0.0078953072 0.3644665449 0.0142061979 0.2878184593  0.07905949
#>     5   82 0.2833973812 0.3250699966 0.2369087118 0.3038892029  0.27082533
#>    92   96 0.4912771849 0.4175778331 0.4149672393 0.5078104701  0.42104749
#>    92   43 0.4912771849 0.4175778331 0.3485513655 0.4359982681  0.41517975
#>    31    5 0.3018658068 0.3365209145 0.2833973812 0.3250699966  0.32231319
#>    47   31 0.3722719635 0.3215140189 0.3018658068 0.3365209145  0.32689109
#>    71   82 0.2723030420 0.2548221585 0.2369087118 0.3038892029  0.23735737
#>    92   14 0.4912771849 0.4175778331 0.5604866282 0.4803359473  0.48725202
#>    14   15 0.5604866282 0.4803359473 0.5633155366 0.4932438849  0.48935370
#>    94   19 0.1019776048 0.2568421801 0.1560813196 0.2211533328  0.14974545
#>    82   45 0.2369087118 0.3038892029 0.1970939521 0.2853621123  0.22824068
#>    81   55 0.6780784277 0.7572867137 0.7612671522 0.7552275941  0.72265287
#>    81   40 0.6780784277 0.7572867137 0.6105767762 0.8078280263  0.62898588
#>    15    7 0.5633155366 0.4932438849 0.7121348998 0.6537576263  0.58711046
#>    19   91 0.1560813196 0.2211533328 0.1460513179 0.1897803601  0.19752121
#>    19   41 0.1560813196 0.2211533328 0.2119771291 0.2400203143  0.17631901
#>    26   71 0.2621041203 0.1462614297 0.2723030420 0.2548221585  0.25332267
#>    37   68 0.0142061979 0.2878184593 0.0002931199 0.0968719184  0.02929662
#>     7    9 0.7121348998 0.6537576263 0.7221879147 0.6382715642  0.76190800
#>    92   76 0.4912771849 0.4175778331 0.5270903010 0.3152210733  0.57862793
#>    92   47 0.4912771849 0.4175778331 0.3722719635 0.3215140189  0.45080322
#>    76   47 0.5270903010 0.3152210733 0.3722719635 0.3215140189  0.45017228
#>    16   47 0.4703617924 0.2546832282 0.3722719635 0.3215140189  0.39690090
#>    91   26 0.1460513179 0.1897803601 0.2621041203 0.1462614297  0.20759390
#>    64   80 0.5722805662 0.8877758082 0.7870738256 0.9865851907  0.65667076
#>     7    6 0.7121348998 0.6537576263 0.7741482164 0.7278018666  0.71695683
#>    47   71 0.3722719635 0.3215140189 0.2723030420 0.2548221585  0.34412407
#>    23   71 0.3641515780 0.1864609084 0.2723030420 0.2548221585  0.30091054
#>     6   55 0.7741482164 0.7278018666 0.7612671522 0.7552275941  0.71873859
#>    55   98 0.7612671522 0.7552275941 0.7901625187 0.7367964922  0.82059894
#>     6    9 0.7741482164 0.7278018666 0.7221879147 0.6382715642  0.79256123
#>    76   32 0.5270903010 0.3152210733 0.5207387908 0.2411128555  0.62546619
#>     9   15 0.7221879147 0.6382715642 0.5633155366 0.4932438849  0.66205952
#>    76   16 0.5270903010 0.3152210733 0.4703617924 0.2546832282  0.50416010
#>    16   23 0.4703617924 0.2546832282 0.3641515780 0.1864609084  0.40551790
#>    23   87 0.3641515780 0.1864609084 0.3832127961 0.1583253913  0.40108079
#>    26   87 0.2621041203 0.1462614297 0.3832127961 0.1583253913  0.32400852
#>     9   22 0.7221879147 0.6382715642 0.7683890950 0.5775087647  0.77759108
#>    32   16 0.5207387908 0.2411128555 0.4703617924 0.2546832282  0.48362020
#>    14   12 0.5604866282 0.4803359473 0.6675236757 0.3694771321  0.61688496
#>    22   15 0.7683890950 0.5775087647 0.5633155366 0.4932438849  0.67671549
#>    28   26 0.0462473714 0.0544427491 0.2621041203 0.1462614297  0.16721096
#>    68   28 0.0002931199 0.0968719184 0.0462473714 0.0544427491  0.07830846
#>    68   91 0.0002931199 0.0968719184 0.1460513179 0.1897803601  0.05225124
#>    70  100 0.0078953072 0.3644665449 0.0000000000 0.5620638226 -3.39821318
#>    79   48 0.7864508736 0.9999378166 0.2697440470 0.9822331923  0.38880576
#>     9   35 0.7221879147 0.6382715642 0.7815064136 0.5768812677  0.79718970
#>    35   22 0.7815064136 0.5768812677 0.7683890950 0.5775087647  0.77038977
#>    55   13 0.7612671522 0.7552275941 0.8902414815 0.8654676848  0.77397325
#>    98   21 0.7901625187 0.7367964922 0.8275979033 0.7211932509  0.83925450
#>    98    6 0.7901625187 0.7367964922 0.7741482164 0.7278018666  0.79828661
#>    73   87 0.3771721839 0.0864272517 0.3832127961 0.1583253913  0.32517816
#>    29   32 0.4997097225 0.1043856225 0.5207387908 0.2411128555  0.50559550
#>    77   20 0.4406626288 0.1121810328 0.4365557469 0.1797528109  0.41961588
#>    73   77 0.3771721839 0.0864272517 0.4406626288 0.1121810328  0.40024186
#>    12   32 0.6675236757 0.3694771321 0.5207387908 0.2411128555  0.62570500
#>    53   79 0.8592558947 1.0000000000 0.7864508736 0.9999378166  0.82285767
#>    65   53 0.8773025085 0.9864624673 0.8592558947 1.0000000000  0.83213175
#>    21   35 0.8275979033 0.7211932509 0.7815064136 0.5768812677  0.84565077
#>    22   49 0.7683890950 0.5775087647 0.6867236185 0.3948962530  0.74462021
#>    49   93 0.6867236185 0.3948962530 0.7327610204 0.3131750715  0.72088225
#>    49   12 0.6867236185 0.3948962530 0.6675236757 0.3694771321  0.71240636
#>    77   29 0.4406626288 0.1121810328 0.4997097225 0.1043856225  0.47545684
#>    13   65 0.8902414815 0.8654676848 0.8773025085 0.9864624673  0.83209828
#>    32   60 0.5207387908 0.2411128555 0.5603967447 0.1170211463  0.61030845
#>     2   21 0.9946118073 0.6858030764 0.8275979033 0.7211932509  0.90367105
#>    50   49 0.7928449846 0.3889727581 0.6867236185 0.3948962530  0.73934463
#>    66   24 0.7335660020 0.2823147963 0.7628203590 0.3357323428  0.76716799
#>    83   27 0.8217832015 0.2958170725 0.8268035563 0.3652857465  0.80333051
#>    29   60 0.4997097225 0.1043856225 0.5603967447 0.1170211463  0.51736341
#>    93   12 0.7327610204 0.3131750715 0.6675236757 0.3694771321  0.66090359
#>    50   46 0.7928449846 0.3889727581 0.9089532669 0.5759793575  0.84459266
#>    24   27 0.7628203590 0.3357323428 0.8268035563 0.3652857465  0.79339531
#>    17   46 0.9506920183 0.4286173291 0.9089532669 0.5759793575  0.85369962
#>    51    4 0.6473611907 0.0971626896 0.6821676763 0.1347244205  0.61618064
#>    66   32 0.7335660020 0.2823147963 0.5207387908 0.2411128555  0.63235985
#>    88   29 0.5917324092 0.0930594561 0.4997097225 0.1043856225  0.53716321
#>    73   26 0.3771721839 0.0864272517 0.2621041203 0.1462614297  0.23217056
#>    83    8 0.8217832015 0.2958170725 0.8556346583 0.3075614052  0.82880878
#>    66   83 0.7335660020 0.2823147963 0.8217832015 0.2958170725  0.77703777
#>    88   51 0.5917324092 0.0930594561 0.6473611907 0.0971626896  0.61505353
#>    29   78 0.4997097225 0.1043856225 0.4254812816 0.0008735065  0.52075537
#>    27    8 0.8268035563 0.3652857465 0.8556346583 0.3075614052  0.90372924
#>    83   75 0.8217832015 0.2958170725 0.7853847247 0.1896878495  0.80499235
#>    66    4 0.7335660020 0.2823147963 0.6821676763 0.1347244205  0.70939808
#>     2   18 0.9946118073 0.6858030764 1.0000000000 0.9182801270  0.98824424
#>    13   21 0.8902414815 0.8654676848 0.8275979033 0.7211932509  0.92413964
#>    17    2 0.9506920183 0.4286173291 0.9946118073 0.6858030764  1.06688089
#>    18   65 1.0000000000 0.9182801270 0.8773025085 0.9864624673  0.92651762
#>    27   17 0.8268035563 0.3652857465 0.9506920183 0.4286173291  0.86121675
#>     8   67 0.8556346583 0.3075614052 0.8608754160 0.2371347266  0.97620088
#>    90   59 0.8483885782 0.0914231212 0.8551043890 0.1361337394  0.78966692
#>    75    4 0.7853847247 0.1896878495 0.6821676763 0.1347244205  0.74697063
#>    88   57 0.5917324092 0.0930594561 0.6159332644 0.0000000000  0.62276701
#>    57   33 0.6159332644 0.0000000000 0.7169038971 0.0799586883  0.66923445
#>    67   97 0.8608754160 0.2371347266 0.8686081559 0.2140302011  0.96077431
#>    67   75 0.8608754160 0.2371347266 0.7853847247 0.1896878495  0.82404174
#>    17    8 0.9506920183 0.4286173291 0.8556346583 0.3075614052  0.99443704
#>    97   75 0.8686081559 0.2140302011 0.7853847247 0.1896878495  0.83338492
#>    33   90 0.7169038971 0.0799586883 0.8483885782 0.0914231212  0.77992860
#>    57   90 0.6159332644 0.0000000000 0.8483885782 0.0914231212  0.80150569
#>    36   10 0.9812294706 0.1577968559 0.9895469812 0.1310468646  1.74721517
#>    97   59 0.8686081559 0.2140302011 0.8551043890 0.1361337394  0.91491756
#>    10    2 0.9895469812 0.1310468646 0.9946118073 0.6858030764  1.89202249
#>    36   59 0.9812294706 0.1577968559 0.8551043890 0.1361337394  0.92198965
#>     1   39 0.0445426497 0.8182835048 0.1256119587 0.8588129805  0.10573396
#>    58   72 0.1681300526 0.7800724814 0.1070682540 0.6871840071  0.09477365
#>    58   61 0.1681300526 0.7800724814 0.2709264618 0.7653392217  0.22136312
#>    72   89 0.1070682540 0.6871840071 0.1797244246 0.6284064139  0.12283816
#>    72   63 0.1070682540 0.6871840071 0.0658449161 0.6304780067  0.05194065
#>    72   42 0.1070682540 0.6871840071 0.0513959401 0.6287016328  0.01174915
#>    39   48 0.1256119587 0.8588129805 0.2697440470 0.9822331923  0.06114313
#>    84   58 0.2330876722 0.9266669517 0.1681300526 0.7800724814  0.20560308
#>    89   69 0.1797244246 0.6284064139 0.1999095636 0.7364870890  0.26926185
#>    52  100 0.1915779281 0.5144500557 0.0000000000 0.5620638226  0.09528925
#>    38   34 0.2592523979 0.9223598056 0.2794968263 0.8003651745  0.23483071
#>    48   38 0.2697440470 0.9822331923 0.2592523979 0.9223598056  0.25112924
#>   100    1 0.0000000000 0.5620638226 0.0445426497 0.8182835048 -0.13197011
#>    38   99 0.2592523979 0.9223598056 0.3688672651 0.9306484457  0.31837077
#>    11   34 0.4875081331 0.8615031451 0.2794968263 0.8003651745  0.38575528
#>    89   61 0.1797244246 0.6284064139 0.2709264618 0.7653392217  0.28148697
#>    52   63 0.1915779281 0.5144500557 0.0658449161 0.6304780067  0.12160815
#>    85   89 0.2829283602 0.5048901793 0.1797244246 0.6284064139  0.28119983
#>    25   52 0.2037826419 0.4429980902 0.1915779281 0.5144500557  0.23467593
#>    34   62 0.2794968263 0.8003651745 0.3828943585 0.6272658180  0.39691663
#>    96   56 0.4149672393 0.5078104701 0.3160067188 0.4538210814  0.34883759
#>    96   85 0.4149672393 0.5078104701 0.2829283602 0.5048901793  0.34789756
#>    34   30 0.2794968263 0.8003651745 0.5135064289 0.8197022944  0.40100666
#>    44   99 0.4733020205 0.9813697625 0.3688672651 0.9306484457  0.37184368
#>    44   11 0.4733020205 0.9813697625 0.4875081331 0.8615031451  0.44019047
#>    85   54 0.2829283602 0.5048901793 0.2806953197 0.4237829532  0.25018777
#>    96   62 0.4149672393 0.5078104701 0.3828943585 0.6272658180  0.40312255
#>    52   74 0.1915779281 0.5144500557 0.0105203711 0.3730006507  0.07923446
#>    56   43 0.3160067188 0.4538210814 0.3485513655 0.4359982681  0.31695642
#>    45   94 0.1970939521 0.2853621123 0.1019776048 0.2568421801  0.12108934
#>    11   64 0.4875081331 0.8615031451 0.5722805662 0.8877758082  0.53398916
#>    40   86 0.6105767762 0.8078280263 0.5812325449 0.7744185289  0.56282594
#>     7   86 0.7121348998 0.6537576263 0.5812325449 0.7744185289  0.61427973
#>    40   30 0.6105767762 0.8078280263 0.5135064289 0.8197022944  0.56463520
#>    96   95 0.4149672393 0.5078104701 0.4613836455 0.5898542784  0.49095267
#>    82   54 0.2369087118 0.3038892029 0.2806953197 0.4237829532  0.23344664
#>    94   37 0.1019776048 0.2568421801 0.0142061979 0.2878184593  0.07905949
#>    82   25 0.2369087118 0.3038892029 0.2037826419 0.4429980902  0.22769994
#>    45   74 0.1970939521 0.2853621123 0.0105203711 0.3730006507  0.12181405
#>     7   95 0.7121348998 0.6537576263 0.4613836455 0.5898542784  0.58131456
#>    47   43 0.3722719635 0.3215140189 0.3485513655 0.4359982681  0.41517975
#>    71    5 0.2723030420 0.2548221585 0.2833973812 0.3250699966  0.32231319
#>    71   31 0.2723030420 0.2548221585 0.3018658068 0.3365209145  0.32689109
#>    41   82 0.2119771291 0.2400203143 0.2369087118 0.3038892029  0.23735737
#>    96   14 0.4149672393 0.5078104701 0.5604866282 0.4803359473  0.48725202
#>    96   15 0.4149672393 0.5078104701 0.5633155366 0.4932438849  0.48935370
#>    45   19 0.1970939521 0.2853621123 0.1560813196 0.2211533328  0.14974545
#>    41   45 0.2119771291 0.2400203143 0.1970939521 0.2853621123  0.22824068
#>    80   55 0.7870738256 0.9865851907 0.7612671522 0.7552275941  0.72265287
#>    41   91 0.2119771291 0.2400203143 0.1460513179 0.1897803601  0.19752121
#>    41   71 0.2119771291 0.2400203143 0.2723030420 0.2548221585  0.25332267
#>    94   68 0.1019776048 0.2568421801 0.0002931199 0.0968719184  0.02929662
#>    49   14 0.6867236185 0.3948962530 0.5604866282 0.4803359473  0.64461745
#>    12   76 0.6675236757 0.3694771321 0.5270903010 0.3152210733  0.57862793
#>    12   92 0.6675236757 0.3694771321 0.4912771849 0.4175778331  0.57864719
#>    23   47 0.3641515780 0.1864609084 0.3722719635 0.3215140189  0.39690090
#>    41   26 0.2119771291 0.2400203143 0.2621041203 0.1462614297  0.20759390
#>    79   80 0.7864508736 0.9999378166 0.7870738256 0.9865851907  0.65667076
#>    40   64 0.6105767762 0.8078280263 0.5722805662 0.8877758082  0.69743416
#>    81    6 0.6780784277 0.7572867137 0.7741482164 0.7278018666  0.71695683
#>    13   98 0.8902414815 0.8654676848 0.7901625187 0.7367964922  0.82059894
#>    21    9 0.8275979033 0.7211932509 0.7221879147 0.6382715642  0.79256123
#>    20   23 0.4365557469 0.1797528109 0.3641515780 0.1864609084  0.40551790
#>    20   87 0.4365557469 0.1797528109 0.3832127961 0.1583253913  0.40108079
#>    20   16 0.4365557469 0.1797528109 0.4703617924 0.2546832282  0.48362020
#>    49   15 0.6867236185 0.3948962530 0.5633155366 0.4932438849  0.67671549
#>    79   64 0.7864508736 0.9999378166 0.5722805662 0.8877758082  0.62668688
#>    80   81 0.7870738256 0.9865851907 0.6780784277 0.7572867137  0.71684717
#>    91   28 0.1460513179 0.1897803601 0.0462473714 0.0544427491  0.07830846
#>    94   91 0.1019776048 0.2568421801 0.1460513179 0.1897803601  0.05225124
#>    68   70 0.0002931199 0.0968719184 0.0078953072 0.3644665449 -0.85283180
#>    68  100 0.0002931199 0.0968719184 0.0000000000 0.5620638226 -3.39821318
#>    44   48 0.4733020205 0.9813697625 0.2697440470 0.9822331923  0.38880576
#>    50   22 0.7928449846 0.3889727581 0.7683890950 0.5775087647  0.77038977
#>    80   13 0.7870738256 0.9865851907 0.8902414815 0.8654676848  0.77397325
#>    21    6 0.8275979033 0.7211932509 0.7741482164 0.7278018666  0.79828661
#>    20   32 0.4365557469 0.1797528109 0.5207387908 0.2411128555  0.50559550
#>    87   77 0.3832127961 0.1583253913 0.4406626288 0.1121810328  0.40024186
#>    78   28 0.4254812816 0.0008735065 0.0462473714 0.0544427491  0.22320843
#>    80   53 0.7870738256 0.9865851907 0.8592558947 1.0000000000  0.83213175
#>    46   35 0.9089532669 0.5759793575 0.7815064136 0.5768812677  0.84565077
#>    24   93 0.7628203590 0.3357323428 0.7327610204 0.3131750715  0.72088225
#>    20   29 0.4365557469 0.1797528109 0.4997097225 0.1043856225  0.47545684
#>    80   65 0.7870738256 0.9865851907 0.8773025085 0.9864624673  0.83209828
#>     4   60 0.6821676763 0.1347244205 0.5603967447 0.1170211463  0.61030845
#>    78   77 0.4254812816 0.0008735065 0.4406626288 0.1121810328  0.46281772
#>    46   21 0.9089532669 0.5759793575 0.8275979033 0.7211932509  0.90367105
#>    24   49 0.7628203590 0.3357323428 0.6867236185 0.3948962530  0.73934463
#>    66   12 0.7335660020 0.2823147963 0.6675236757 0.3694771321  0.66090359
#>    50   27 0.7928449846 0.3889727581 0.8268035563 0.3652857465  0.79339531
#>     4   32 0.6821676763 0.1347244205 0.5207387908 0.2411128555  0.63235985
#>    57   29 0.6159332644 0.0000000000 0.4997097225 0.1043856225  0.53716321
#>    78   73 0.4254812816 0.0008735065 0.3771721839 0.0864272517  0.42587068
#>    78   26 0.4254812816 0.0008735065 0.2621041203 0.1462614297  0.23217056
#>    24   83 0.7628203590 0.3357323428 0.8217832015 0.2958170725  0.77703777
#>    75   66 0.7853847247 0.1896878495 0.7335660020 0.2823147963  0.78372052
#>    60   51 0.5603967447 0.1170211463 0.6473611907 0.0971626896  0.61505353
#>    57   78 0.6159332644 0.0000000000 0.4254812816 0.0008735065  0.52075537
#>    13   18 0.8902414815 0.8654676848 1.0000000000 0.9182801270  0.98824424
#>    46    2 0.9089532669 0.5759793575 0.9946118073 0.6858030764  1.06688089
#>    33   51 0.7169038971 0.0799586883 0.6473611907 0.0971626896  0.68440443
#>    50   17 0.7928449846 0.3889727581 0.9506920183 0.4286173291  0.86121675
#>    36   67 0.9812294706 0.1577968559 0.8608754160 0.2371347266  0.97620088
#>    67   83 0.8608754160 0.2371347266 0.8217832015 0.2958170725  0.84912421
#>    75   59 0.7853847247 0.1896878495 0.8551043890 0.1361337394  0.78966692
#>    33    4 0.7169038971 0.0799586883 0.6821676763 0.1347244205  0.74697063
#>    51   57 0.6473611907 0.0971626896 0.6159332644 0.0000000000  0.62276701
#>    36   97 0.9812294706 0.1577968559 0.8686081559 0.2140302011  0.96077431
#>    36    8 0.9812294706 0.1577968559 0.8556346583 0.3075614052  0.99443704
#>    75   90 0.7853847247 0.1896878495 0.8483885782 0.0914231212  0.77992860
#>    10   90 0.9895469812 0.1310468646 0.8483885782 0.0914231212  0.92118145
#>    17   10 0.9506920183 0.4286173291 0.9895469812 0.1310468646  1.74721517
#>    10   59 0.9895469812 0.1310468646 0.8551043890 0.1361337394  0.92198965
#>    10   57 0.9895469812 0.1310468646 0.6159332644 0.0000000000  1.44895797
#>     1   58 0.0445426497 0.8182835048 0.1681300526 0.7800724814  0.10573396
#>    58   69 0.1681300526 0.7800724814 0.1999095636 0.7364870890  0.22136312
#>    39   84 0.1256119587 0.8588129805 0.2330876722 0.9266669517  0.06114313
#>    25   85 0.2037826419 0.4429980902 0.2829283602 0.5048901793  0.23467593
#>    95   30 0.4613836455 0.5898542784 0.5135064289 0.8197022944  0.46989764
#>    74   70 0.0105203711 0.3730006507 0.0078953072 0.3644665449  0.09217997
#>     7   81 0.7121348998 0.6537576263 0.6780784277 0.7572867137  0.61427973
#>    40   80 0.6105767762 0.8078280263 0.7870738256 0.9865851907  0.69743416
#>    26   23 0.2621041203 0.1462614297 0.3641515780 0.1864609084  0.30091054
#>    79   44 0.7864508736 0.9999378166 0.4733020205 0.9813697625  0.62668688
#>    50   35 0.7928449846 0.3889727581 0.7815064136 0.5768812677  0.77038977
#>    24   50 0.7628203590 0.3357323428 0.7928449846 0.3889727581  0.73934463
#>    93   66 0.7327610204 0.3131750715 0.7335660020 0.2823147963  0.76716799
#>    88   60 0.5917324092 0.0930594561 0.5603967447 0.1170211463  0.54082691
#>    13    2 0.8902414815 0.8654676848 0.9946118073 0.6858030764  0.98824424
#>    18   53 1.0000000000 0.9182801270 0.8592558947 1.0000000000  1.22735507
#>    33   75 0.7169038971 0.0799586883 0.7853847247 0.1896878495  0.74697063
#>    36   17 0.9812294706 0.1577968559 0.9506920183 0.4286173291  0.99443704
#>    10   18 0.9895469812 0.1310468646 1.0000000000 0.9182801270 29.02113284
#>          my1          mx2          my2 bp1 bp2
#>   0.79722965   0.20560308   0.85115673   0   0
#>   0.73993259   0.16104756   0.67961425   0   0
#>   0.76178031   0.01174915   0.72218323   0   0
#>   0.78550799   0.26926185   0.66760964   0   0
#>   0.75625110   0.12800885   0.73993259   0   0
#>   0.63238283   0.12160815   0.56476661   0   0
#>   0.84773114   0.22281984   0.79567175   0   0
#>   0.67961425   0.21687619   0.67739318   0   0
#>   0.67739318   0.18123748   0.75625110   0   0
#>   0.68392292   0.06671519   0.56374726   0   0
#>   0.79567175   0.39691663   0.75307268   0   0
#>   0.72218323  -0.13197011   0.71698785   0   0
#>   1.07997068   0.25112924   0.95463915   0   0
#>   0.53624629   0.06671519   0.56374726   0   0
#>   0.85563009   0.21333384   0.84773114   0   0
#>   1.35149597  -2.28424902   4.25375650   0   1
#>   0.95463915   0.23483071   0.85563009   0   0
#>   0.94384151   0.37184368   1.05739582   0   0
#>   0.56374726  -0.13197011   0.71698785   0   0
#>   0.86949317   0.38575528   0.82326941   0   0
#>   0.65946728   0.39691663   0.75307268   0   0
#>   0.57753450   0.12160815   0.56476661   0   0
#>   0.77758353   0.53398916   0.86142698   0   0
#>   0.60832002   0.28148697   0.65946728   0   0
#>   0.48504330   0.24435519   0.57753450   0   0
#>   0.46004125   0.12181405   0.36751623   0   0
#>   0.51133348   0.27650542   0.46448266   0   0
#>   0.55383610   0.28119983   0.60832002   0   0
#>   0.46448266   0.31695642   0.41693040   0   0
#>   0.75551578   0.46989764   0.70875752   0   0
#>   0.91667038   0.38575528   0.82326941   0   0
#>   0.92543595   0.62668688   1.04444563   0   0
#>   0.46520724   0.22769994   0.37519492   0   0
#>   0.56866360   0.46989764   0.70875752   0   0
#>   0.49320524   0.42104749   0.43556849   0   0
#>   0.34321163   0.07905949   0.33174206   0   0
#>   0.47164852  -0.26338845   0.45258335   0   0
#>   0.41693040   0.32222571   0.38765948   0   0
#>   0.36486414   0.12181405   0.36751623   0   0
#>   0.36597339   0.09217997   0.34321163   0   0
#>   0.86142698   0.56463520   0.83496744   0   0
#>   0.82017698   0.48590094   0.70512844   0   0
#>   0.70512844   0.58131456   0.64317052   0   0
#>   0.67893370   0.62898588   0.76206738   0   0
#>   0.37599427   0.32222571   0.38765948   0   0
#>   0.37398324   0.32222571   0.38765948   0   0
#>   0.51897358   0.58711046   0.62042793   0   0
#>   0.37309616   0.26585401   0.37398324   0   0
#>   0.33174206  -0.85283180   0.25501394   0   0
#>   0.29105553   0.23344664   0.37309616   0   0
#>   0.43556849   0.48725202   0.49155781   0   0
#>   0.39010383   0.42104749   0.43556849   0   0
#>   0.28292398   0.26585401   0.37398324   0   0
#>   0.28126746   0.34708187   0.37599427   0   0
#>   0.26691353   0.27082533   0.29105553   0   0
#>   0.49155781   0.57864719   0.39076751   0   0
#>   0.50268943   0.64461745   0.46866177   0   0
#>   0.27040286   0.11460124   0.21712480   0   0
#>   0.27047228   0.18431710   0.36486414   0   0
#>   0.87665290   0.71873859   0.71851538   0   0
#>   0.76206738   0.71684717   0.87941260   0   0
#>   0.62042793   0.62443683   0.58582101   0   0
#>   0.19061513   0.11460124   0.21712480   0   0
#>   0.25342930   0.19752121   0.19061513   0   0
#>   0.20184586   0.30091054   0.19737514   0   0
#>   0.19073877  -0.85283180   0.25501394   0   0
#>   0.67506253   0.62443683   0.58582101   0   0
#>   0.39069695   0.45080322   0.34597297   0   0
#>   0.34597297   0.41517975   0.39010383   0   0
#>   0.33045068   0.45080322   0.34597297   0   0
#>   0.25226246   0.45017228   0.33045068   0   0
#>   0.17739755   0.16721096   0.06970757   0   0
#>   0.98719222   0.69743416   0.89858017   0   0
#>   0.71270990   0.76190800   0.67506253   0   0
#>   0.25543579   0.32689109   0.28126746   0   0
#>   0.19737514   0.34412407   0.25543579   0   0
#>   0.71851538   0.77504404   0.74496040   0   0
#>   0.81637912   0.77504404   0.74496040   0   0
#>   0.65727245   0.76190800   0.67506253   0   0
#>   0.26946339   0.50416010   0.27986004   0   0
#>   0.54460676   0.62443683   0.58582101   0   0
#>   0.27986004   0.45017228   0.33045068   0   0
#>   0.23884730   0.39690090   0.25226246   0   0
#>   0.19095512   0.32400852   0.13874029   0   0
#>   0.13874029   0.32517816   0.12699842   0   0
#>   0.63245152   0.66205952   0.54460676   0   0
#>   0.20361021   0.50416010   0.27986004   0   0
#>   0.42768706   0.57864719   0.39076751   0   0
#>   0.50893887   0.66205952   0.54460676   0   0
#>   0.06970757   0.22320843  -0.06193702   0   0
#>   0.13526821 -10.76751710 -11.61164561   0   1
#>   0.17614775   0.07830846   0.13526821   0   0
#>   0.32732655  -0.26338845   0.45258335   0   0
#>   5.05629297   0.32257058   6.98935691   0   1
#>   0.65138874   0.77759108   0.63245152   0   0
#>   0.48191413   0.77759108   0.63245152   0   0
#>   0.87092839   0.82059894   0.81637912   0   0
#>   0.80186904   0.79828661   0.70357864   0   0
#>   0.70357864   0.77504404   0.74496040   0   0
#>   0.12699842   0.40024186   0.12069185   0   0
#>   0.17346116   0.51736341   0.17165122   0   0
#>   0.14481255   0.47545684   0.14820645   0   0
#>   0.12069185   0.42587068   0.05750946   0   0
#>   0.26919031   0.62546619   0.26946339   0   0
#>   0.99494549   0.81112624  14.73025186   0   1
#>   0.94504379   1.22735507   1.47190808   0   0
#>   0.63591086   0.79718970   0.65138874   0   0
#>   0.47857144   0.67671549   0.50893887   0   0
#>   0.36031131   0.71240636   0.35553644   0   0
#>   0.35553644   0.61688496   0.42768706   0   0
#>   0.14820645   0.46281772   0.05247021   0   0
#>   0.92043918   0.92651762   0.93053622   0   0
#>   0.20135514   0.51736341   0.17165122   0   0
#>   0.66841644   0.92413964   0.76501214   0   0
#>   0.38405766   0.74462021   0.47857144   0   0
#>   0.29863193   0.77703777   0.29322670   0   0
#>   0.33206635   0.82880878   0.33022509   0   0
#>   0.17165122   0.54082691   0.05895859   0   0
#>   0.29586005   0.71240636   0.35553644   0   0
#>   0.48639160   0.85369962   0.48073729   0   0
#>   0.35357609   0.80333051   0.33206635   0   0
#>   0.48073729   1.06688089   0.54111866   0   0
#>   0.16096360   0.68440443   0.09774420   0   0
#>   0.23481491   0.62570500   0.26919031   0   0
#>   0.02919176   0.54082691   0.05895859   0   0
#>  -0.05186598   0.32517816   0.12699842   0   0
#>   0.33022509   0.84912421   0.27166860   0   0
#>   0.29322670   0.78372052   0.24956490   0   0
#>   0.15602772   0.62276701   0.05145371   0   0
#>   0.01092318   0.46281772   0.05247021   0   0
#>   0.36764500   0.82880878   0.33022509   0   0
#>   0.24226943   0.78372052   0.24956490   0   0
#>   0.20798635   0.63235985   0.23481491   0   0
#>   0.80225163  29.02113284   0.15252467   0   0
#>   0.76501214   0.83925450   0.80186904   0   0
#>   0.54111866   1.89202249   0.40020865   0   0
#>   0.93053622   1.22735507   1.47190808   0   0
#>   0.45080741   0.90372924   0.36764500   0   0
#>   0.28112493   0.84912421   0.27166860   0   0
#>   0.12310316   0.92118145   0.10334887   0   0
#>   0.13742802   0.70939808   0.20798635   0   0
#>   0.05145371   0.53716321   0.02919176   0   0
#>   0.03642350   0.80150569  -0.13060663   0   0
#>   0.25772312   0.82404174   0.21196077   0   0
#>   0.21196077   0.80499235   0.24226943   0   0
#>   0.29641805   0.90372924   0.36764500   0   0
#>   0.18001763   0.82404174   0.21196077   0   0
#>   0.11685929   0.80150569  -0.13060663   0   0
#>  -0.13060663   1.44895797  -1.77683925   0   0
#>   0.38130058   0.92198965   0.12470902   0   0
#>   0.16588351   0.83338492   0.18001763   0   0
#>   0.40020865  29.02113284   0.15252467   0   0
#>   0.12470902   0.91491756   0.16588351   0   0
#>   0.79722965  -0.17136379   1.35149597   0   0
#>   0.76178031   0.12800885   0.73993259   0   0
#>   0.78550799   0.22281984   0.79567175   0   0
#>   0.63238283   0.16104756   0.67961425   0   0
#>   0.68392292   0.12283816   0.63238283   0   0
#>   0.72218323   0.05194065   0.68392292   0   0
#>   1.07997068  -0.17136379   1.35149597   0   0
#>   0.85115673   0.21333384   0.84773114   0   0
#>   0.66760964   0.21687619   0.67739318   0   0
#>   0.53624629   0.07923446   0.47164852   0   0
#>   0.85563009   0.31837077   0.86949317   0   0
#>   0.95463915   0.31274885   0.94384151   0   0
#>   0.71698785  -3.92037664   1.37558544   0   1
#>   0.86949317   0.31274885   0.94384151   0   0
#>   0.82326941   0.39918311   0.77758353   0   0
#>   0.65946728   0.26926185   0.66760964   0   0
#>   0.56476661   0.09528925   0.53624629   0   0
#>   0.60832002   0.24435519   0.57753450   0   0
#>   0.48504330   0.08830252   0.46004125   0   0
#>   0.75307268   0.40100666   0.75551578   0   0
#>   0.51133348   0.35872772   0.49320524   0   0
#>   0.55383610   0.34883759   0.51133348   0   0
#>   0.75551578   0.39918311   0.77758353   0   0
#>   1.05739582   0.44019047   0.91667038   0   0
#>   0.91667038   0.51415151   0.92543595   0   0
#>   0.46520724   0.27650542   0.46448266   0   0
#>   0.56866360   0.34789756   0.55383610   0   0
#>   0.47164852   0.08830252   0.46004125   0   0
#>   0.41693040   0.35872772   0.49320524   0   0
#>   0.36597339   0.14974545   0.27040286   0   0
#>   0.86142698   0.51415151   0.92543595   0   0
#>   0.82017698   0.62898588   0.76206738   0   0
#>   0.67893370   0.58131456   0.64317052   0   0
#>   0.83496744   0.56282594   0.82017698   0   0
#>   0.51897358   0.40312255   0.56866360   0   0
#>   0.37309616   0.22769994   0.37519492   0   0
#>   0.33174206   0.02929662   0.19073877   0   0
#>   0.37519492   0.18431710   0.36486414   0   0
#>   0.36751623   0.12108934   0.36597339   0   0
#>   0.64317052   0.58711046   0.62042793   0   0
#>   0.39010383   0.34708187   0.37599427   0   0
#>   0.28292398   0.27082533   0.29105553   0   0
#>   0.28126746   0.32231319   0.28292398   0   0
#>   0.26691353   0.22824068   0.27047228   0   0
#>   0.49155781   0.48935370   0.50268943   0   0
#>   0.50268943   0.49095267   0.51897358   0   0
#>   0.27040286   0.17631901   0.25342930   0   0
#>   0.27047228   0.17631901   0.25342930   0   0
#>   0.87665290   0.77397325   0.87092839   0   0
#>   0.19061513   0.20759390   0.17739755   0   0
#>   0.20184586   0.23735737   0.26691353   0   0
#>   0.19073877   0.05225124   0.17614775   0   0
#>   0.46866177   0.61688496   0.42768706   0   0
#>   0.39069695   0.62546619   0.26946339   0   0
#>   0.39076751   0.57862793   0.39069695   0   0
#>   0.25226246   0.34412407   0.25543579   0   0
#>   0.17739755   0.25332267   0.20184586   0   0
#>   0.98719222   0.82285767   0.99494549   0   0
#>   0.89858017   0.56463520   0.83496744   0   0
#>   0.71270990   0.71873859   0.71851538   0   0
#>   0.81637912   0.83925450   0.80186904   0   0
#>   0.65727245   0.79718970   0.65138874   0   0
#>   0.23884730   0.40108079   0.19095512   0   0
#>   0.19095512   0.41961588   0.14481255   0   0
#>   0.20361021   0.40551790   0.23884730   0   0
#>   0.50893887   0.64461745   0.46866177   0   0
#>   1.04444563   0.65667076   0.98719222   0   0
#>   0.87941260   0.72265287   0.87665290   0   0
#>   0.13526821   0.16721096   0.06970757   0   0
#>   0.17614775   0.11460124   0.21712480   0   0
#>   0.25501394  -3.39821318   0.32732655   0   0
#>   0.32732655  -5.54786280   0.32597204   0   1
#>   5.05629297   0.37184368   1.05739582   0   0
#>   0.48191413   0.74462021   0.47857144   0   0
#>   0.87092839   0.83209828   0.92043918   0   0
#>   0.70357864   0.79256123   0.65727245   0   0
#>   0.17346116   0.48362020   0.20361021   0   0
#>   0.12069185   0.41961588   0.14481255   0   0
#>  -0.06193702  -0.14198323  -2.64724633   0   1
#>   0.94504379   0.82285767   0.99494549   0   0
#>   0.63591086   0.84459266   0.48639160   0   0
#>   0.36031131   0.76716799   0.29863193   0   0
#>   0.14820645   0.50559550   0.17346116   0   0
#>   0.92043918   0.83213175   0.94504379   0   0
#>   0.20135514   0.61618064   0.16096360   0   0
#>   0.05247021   0.42587068   0.05750946   0   0
#>   0.66841644   0.84565077   0.63591086   0   0
#>   0.38405766   0.72088225   0.36031131   0   0
#>   0.29586005   0.62570500   0.26919031   0   0
#>   0.35357609   0.86121675   0.45080741   0   0
#>   0.23481491   0.61030845   0.20135514   0   0
#>   0.02919176   0.52075537   0.01092318   0   0
#>   0.05750946   0.23217056  -0.05186598   0   0
#>  -0.05186598   0.22320843  -0.06193702   0   0
#>   0.29322670   0.80333051   0.33206635   0   0
#>   0.24956490   0.70939808   0.20798635   0   0
#>   0.15602772   0.61618064   0.16096360   0   0
#>   0.01092318   0.49667373  -5.23963368   0   1
#>   0.80225163   0.92651762   0.93053622   0   0
#>   0.54111866   0.90367105   0.66841644   0   0
#>   0.09774420   0.66923445   0.03642350   0   0
#>   0.45080741   0.85369962   0.48073729   0   0
#>   0.28112493   0.96077431   0.25772312   0   0
#>   0.27166860   0.80499235   0.24226943   0   0
#>   0.12310316   0.83338492   0.18001763   0   0
#>   0.13742802   0.68440443   0.09774420   0   0
#>   0.05145371   0.66923445   0.03642350   0   0
#>   0.25772312   0.91491756   0.16588351   0   0
#>   0.29641805   0.97620088   0.28112493   0   0
#>   0.11685929   0.78966692   0.12310316   0   0
#>   0.10334887   1.44895797  -1.77683925   0   0
#>   0.38130058   1.89202249   0.40020865   0   0
#>   0.12470902   0.92118145   0.10334887   0   0
#>  -1.77683925   2.28492668  -4.16018033   0   1
#>   0.79722965   0.09477365   0.76178031   0   0
#>   0.78550799   0.18123748   0.75625110   0   0
#>   1.07997068   0.20560308   0.85115673   0   0
#>   0.48504330   0.25018777   0.46520724   0   0
#>   0.70875752   0.48590094   0.70512844   0   0
#>   0.34321163  -0.26338845   0.45258335   0   0
#>   0.67893370   0.71695683   0.71270990   0   0
#>   0.89858017   0.71684717   0.87941260   0   0
#>   0.19737514   0.32400852   0.13874029   0   0
#>   1.04444563   0.38880576   5.05629297   0   0
#>   0.48191413   0.84459266   0.48639160   0   0
#>   0.38405766   0.79339531   0.35357609   0   0
#>   0.29863193   0.66090359   0.29586005   0   0
#>   0.05895859   0.61505353   0.15602772   0   0
#>   0.80225163   0.92413964   0.76501214   0   0
#>   1.47190808   4.31263264   6.78560493   0   1
#>   0.13742802   0.77992860   0.11685929   0   0
#>   0.29641805   1.74721517   0.38130058   0   0
#>   0.15252467  30.29118045   0.13566076   0   1
#> 
#> $x
#>                [,1]         [,2]
#>   [1,] 0.0445426497 0.8182835048
#>   [2,] 0.9946118073 0.6858030764
#>   [3,] 0.1837334308 0.7291765178
#>   [4,] 0.6821676763 0.1347244205
#>   [5,] 0.2833973812 0.3250699966
#>   [6,] 0.7741482164 0.7278018666
#>   [7,] 0.7121348998 0.6537576263
#>   [8,] 0.8556346583 0.3075614052
#>   [9,] 0.7221879147 0.6382715642
#>  [10,] 0.9895469812 0.1310468646
#>  [11,] 0.4875081331 0.8615031451
#>  [12,] 0.6675236757 0.3694771321
#>  [13,] 0.8902414815 0.8654676848
#>  [14,] 0.5604866282 0.4803359473
#>  [15,] 0.5633155366 0.4932438849
#>  [16,] 0.4703617924 0.2546832282
#>  [17,] 0.9506920183 0.4286173291
#>  [18,] 1.0000000000 0.9182801270
#>  [19,] 0.1560813196 0.2211533328
#>  [20,] 0.4365557469 0.1797528109
#>  [21,] 0.8275979033 0.7211932509
#>  [22,] 0.7683890950 0.5775087647
#>  [23,] 0.3641515780 0.1864609084
#>  [24,] 0.7628203590 0.3357323428
#>  [25,] 0.2037826419 0.4429980902
#>  [26,] 0.2621041203 0.1462614297
#>  [27,] 0.8268035563 0.3652857465
#>  [28,] 0.0462473714 0.0544427491
#>  [29,] 0.4997097225 0.1043856225
#>  [30,] 0.5135064289 0.8197022944
#>  [31,] 0.3018658068 0.3365209145
#>  [32,] 0.5207387908 0.2411128555
#>  [33,] 0.7169038971 0.0799586883
#>  [34,] 0.2794968263 0.8003651745
#>  [35,] 0.7815064136 0.5768812677
#>  [36,] 0.9812294706 0.1577968559
#>  [37,] 0.0142061979 0.2878184593
#>  [38,] 0.2592523979 0.9223598056
#>  [39,] 0.1256119587 0.8588129805
#>  [40,] 0.6105767762 0.8078280263
#>  [41,] 0.2119771291 0.2400203143
#>  [42,] 0.0513959401 0.6287016328
#>  [43,] 0.3485513655 0.4359982681
#>  [44,] 0.4733020205 0.9813697625
#>  [45,] 0.1970939521 0.2853621123
#>  [46,] 0.9089532669 0.5759793575
#>  [47,] 0.3722719635 0.3215140189
#>  [48,] 0.2697440470 0.9822331923
#>  [49,] 0.6867236185 0.3948962530
#>  [50,] 0.7928449846 0.3889727581
#>  [51,] 0.6473611907 0.0971626896
#>  [52,] 0.1915779281 0.5144500557
#>  [53,] 0.8592558947 1.0000000000
#>  [54,] 0.2806953197 0.4237829532
#>  [55,] 0.7612671522 0.7552275941
#>  [56,] 0.3160067188 0.4538210814
#>  [57,] 0.6159332644 0.0000000000
#>  [58,] 0.1681300526 0.7800724814
#>  [59,] 0.8551043890 0.1361337394
#>  [60,] 0.5603967447 0.1170211463
#>  [61,] 0.2709264618 0.7653392217
#>  [62,] 0.3828943585 0.6272658180
#>  [63,] 0.0658449161 0.6304780067
#>  [64,] 0.5722805662 0.8877758082
#>  [65,] 0.8773025085 0.9864624673
#>  [66,] 0.7335660020 0.2823147963
#>  [67,] 0.8608754160 0.2371347266
#>  [68,] 0.0002931199 0.0968719184
#>  [69,] 0.1999095636 0.7364870890
#>  [70,] 0.0078953072 0.3644665449
#>  [71,] 0.2723030420 0.2548221585
#>  [72,] 0.1070682540 0.6871840071
#>  [73,] 0.3771721839 0.0864272517
#>  [74,] 0.0105203711 0.3730006507
#>  [75,] 0.7853847247 0.1896878495
#>  [76,] 0.5270903010 0.3152210733
#>  [77,] 0.4406626288 0.1121810328
#>  [78,] 0.4254812816 0.0008735065
#>  [79,] 0.7864508736 0.9999378166
#>  [80,] 0.7870738256 0.9865851907
#>  [81,] 0.6780784277 0.7572867137
#>  [82,] 0.2369087118 0.3038892029
#>  [83,] 0.8217832015 0.2958170725
#>  [84,] 0.2330876722 0.9266669517
#>  [85,] 0.2829283602 0.5048901793
#>  [86,] 0.5812325449 0.7744185289
#>  [87,] 0.3832127961 0.1583253913
#>  [88,] 0.5917324092 0.0930594561
#>  [89,] 0.1797244246 0.6284064139
#>  [90,] 0.8483885782 0.0914231212
#>  [91,] 0.1460513179 0.1897803601
#>  [92,] 0.4912771849 0.4175778331
#>  [93,] 0.7327610204 0.3131750715
#>  [94,] 0.1019776048 0.2568421801
#>  [95,] 0.4613836455 0.5898542784
#>  [96,] 0.4149672393 0.5078104701
#>  [97,] 0.8686081559 0.2140302011
#>  [98,] 0.7901625187 0.7367964922
#>  [99,] 0.3688672651 0.9306484457
#> [100,] 0.0000000000 0.5620638226
#> 
#> $tri.obj
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,58,39), [14,18,3]
#> 2: (58,72,3), [8,5,3]
#> 3: (58,1,72), [12,2,1]
#> 4: (58,69,61), [15,11,5]
#> 5: (69,58,3), [2,9,4]
#> 6: (72,63,89), [27,8,10]
#> 7: (84,58,34), [11,17,14]
#> 8: (72,89,3), [9,2,6]
#> 9: (89,69,3), [5,8,15]
#> 10: (72,42,63), [21,6,12]
#> 11: (58,61,34), [32,7,4]
#> 12: (72,1,42), [22,10,3]
#> 13: (39,84,48), [19,18,14]
#> 14: (84,39,58), [1,7,13]
#> 15: (89,61,69), [4,9,25]
#> 16: (52,63,100), [21,45,27]
#> 17: (38,84,34), [7,23,19]
#> 18: (39,48,1), [0,1,13]
#> 19: (48,84,38), [17,20,13]
#> 20: (38,99,48), [37,19,23]
#> 21: (63,42,100), [22,16,10]
#> 22: (100,42,1), [12,0,21]
#> 23: (38,34,99), [24,20,17]
#> 24: (11,99,34), [23,28,38]
#> 25: (89,62,61), [32,15,29]
#> 26: (85,89,52), [27,30,29]
#> 27: (52,89,63), [6,16,26]
#> 28: (34,30,11), [50,24,36]
#> 29: (85,62,89), [25,26,34]
#> 30: (25,85,52), [26,31,41]
#> 31: (52,74,25), [63,30,45]
#> 32: (34,61,62), [25,36,11]
#> 33: (96,85,56), [35,43,34]
#> 34: (96,62,85), [29,33,42]
#> 35: (85,54,56), [46,33,41]
#> 36: (34,62,30), [40,28,32]
#> 37: (44,48,99), [20,38,117]
#> 38: (44,99,11), [24,39,37]
#> 39: (11,64,44), [110,38,50]
#> 40: (95,30,62), [36,42,52]
#> 41: (85,25,54), [62,35,30]
#> 42: (96,95,62), [40,34,58]
#> 43: (56,43,96), [65,33,46]
#> 44: (74,70,94), [60,48,49]
#> 45: (52,100,74), [49,31,16]
#> 46: (56,54,43), [55,43,35]
#> 47: (82,25,45), [63,73,62]
#> 48: (45,74,94), [44,72,63]
#> 49: (70,74,100), [45,116,44]
#> 50: (11,30,64), [57,39,28]
#> 51: (40,30,86), [52,75,57]
#> 52: (30,95,86), [64,51,40]
#> 53: (7,81,86), [75,64,93]
#> 54: (47,43,31), [55,68,66]
#> 55: (31,43,54), [46,56,54]
#> 56: (5,31,54), [55,59,67]
#> 57: (40,64,30), [50,51,92]
#> 58: (96,15,95), [76,42,71]
#> 59: (82,5,54), [56,62,61]
#> 60: (94,70,37), [115,81,44]
#> 61: (71,5,82), [59,69,67]
#> 62: (82,54,25), [41,47,59]
#> 63: (45,25,74), [31,48,47]
#> 64: (7,86,95), [52,76,53]
#> 65: (43,92,96), [70,43,66]
#> 66: (47,92,43), [65,54,87]
#> 67: (71,31,5), [56,61,68]
#> 68: (71,47,31), [54,67,94]
#> 69: (41,71,82), [61,73,80]
#> 70: (96,92,14), [86,71,65]
#> 71: (96,14,15), [82,58,70]
#> 72: (45,94,19), [79,78,48]
#> 73: (41,82,45), [47,78,69]
#> 74: (80,81,55), [96,120,111]
#> 75: (86,81,40), [111,51,53]
#> 76: (95,15,7), [84,64,58]
#> 77: (41,19,91), [79,90,78]
#> 78: (45,19,41), [77,73,72]
#> 79: (94,91,19), [77,72,114]
#> 80: (41,26,71), [95,69,90]
#> 81: (94,37,68), [115,114,60]
#> 82: (49,15,14), [71,108,109]
#> 83: (6,7,9), [84,99,93]
#> 84: (15,9,7), [83,76,101]
#> 85: (12,92,76), [87,100,86]
#> 86: (12,14,92), [70,85,108]
#> 87: (76,92,47), [66,88,85]
#> 88: (16,76,47), [87,89,102]
#> 89: (23,16,47), [88,94,103]
#> 90: (41,91,26), [112,80,77]
#> 91: (79,64,80), [92,129,110]
#> 92: (40,80,64), [91,57,111]
#> 93: (81,7,6), [83,96,53]
#> 94: (23,47,71), [68,95,89]
#> 95: (26,23,71), [94,80,105]
#> 96: (81,6,55), [98,74,93]
#> 97: (13,55,98), [98,121,120]
#> 98: (98,55,6), [96,122,97]
#> 99: (21,6,9), [83,118,122]
#> 100: (12,76,32), [102,128,85]
#> 101: (22,9,15), [84,109,106]
#> 102: (32,76,16), [88,107,100]
#> 103: (20,16,23), [89,104,107]
#> 104: (20,23,87), [105,125,103]
#> 105: (23,26,87), [123,104,95]
#> 106: (35,9,22), [101,119,118]
#> 107: (20,32,16), [102,103,124]
#> 108: (49,14,12), [86,134,82]
#> 109: (49,22,15), [101,82,132]
#> 110: (79,44,64), [39,91,117]
#> 111: (80,40,81), [75,74,92]
#> 112: (91,28,26), [127,90,113]
#> 113: (91,68,28), [0,112,114]
#> 114: (94,68,91), [113,79,81]
#> 115: (68,37,70), [60,116,81]
#> 116: (68,70,100), [49,0,115]
#> 117: (44,79,48), [0,37,110]
#> 118: (21,9,35), [106,131,99]
#> 119: (50,35,22), [106,132,145]
#> 120: (80,55,13), [97,136,74]
#> 121: (13,98,21), [122,163,97]
#> 122: (21,98,6), [98,99,121]
#> 123: (26,73,87), [126,105,153]
#> 124: (20,29,32), [143,107,135]
#> 125: (87,77,20), [135,104,126]
#> 126: (87,73,77), [152,125,123]
#> 127: (78,26,28), [112,0,153]
#> 128: (66,12,32), [100,149,144]
#> 129: (80,53,79), [0,91,130]
#> 130: (80,65,53), [165,129,136]
#> 131: (46,21,35), [118,145,139]
#> 132: (50,22,49), [109,140,119]
#> 133: (24,49,93), [134,141,140]
#> 134: (93,49,12), [108,144,133]
#> 135: (20,77,29), [138,124,125]
#> 136: (80,13,65), [166,130,120]
#> 137: (4,32,60), [143,148,149]
#> 138: (78,29,77), [135,152,158]
#> 139: (46,2,21), [163,131,164]
#> 140: (24,50,49), [132,133,146]
#> 141: (93,66,24), [155,133,144]
#> 142: (24,83,27), [154,146,155]
#> 143: (32,29,60), [150,137,124]
#> 144: (66,93,12), [134,128,141]
#> 145: (35,50,46), [147,131,119]
#> 146: (50,24,27), [142,168,140]
#> 147: (50,17,46), [164,145,168]
#> 148: (60,51,4), [167,137,157]
#> 149: (4,66,32), [128,137,161]
#> 150: (88,60,29), [143,151,157]
#> 151: (57,88,29), [150,158,173]
#> 152: (78,77,73), [126,153,138]
#> 153: (78,73,26), [123,127,152]
#> 154: (27,83,8), [170,159,142]
#> 155: (24,66,83), [156,142,141]
#> 156: (75,83,66), [155,161,160]
#> 157: (60,88,51), [173,148,150]
#> 158: (57,29,78), [138,0,151]
#> 159: (17,27,8), [154,177,168]
#> 160: (67,83,75), [156,176,170]
#> 161: (75,66,4), [149,172,156]
#> 162: (13,2,18), [184,166,163]
#> 163: (2,13,21), [121,139,162]
#> 164: (46,17,2), [185,139,147]
#> 165: (18,53,65), [130,166,0]
#> 166: (13,18,65), [165,136,162]
#> 167: (33,4,51), [148,174,172]
#> 168: (50,27,17), [159,147,146]
#> 169: (36,8,67), [170,175,177]
#> 170: (67,8,83), [154,160,169]
#> 171: (75,90,59), [180,178,179]
#> 172: (33,75,4), [161,167,179]
#> 173: (51,88,57), [151,174,157]
#> 174: (51,57,33), [181,167,173]
#> 175: (36,67,97), [176,183,169]
#> 176: (97,67,75), [160,178,175]
#> 177: (36,17,8), [159,169,182]
#> 178: (59,97,75), [176,171,183]
#> 179: (75,33,90), [181,171,172]
#> 180: (10,59,90), [171,187,186]
#> 181: (33,57,90), [187,179,174]
#> 182: (17,36,10), [186,185,177]
#> 183: (36,97,59), [178,186,175]
#> 184: (10,18,2), [162,185,0]
#> 185: (17,10,2), [184,164,182]
#> 186: (10,36,59), [183,180,182]
#> 187: (10,90,57), [181,0,180]
#> boundary nodes:  48 1 100 68 28 78 57 10 18 53 79 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.089486616 0.087412358 0.145246444 0.076654063 0.053234054 0.113898349 
#>                                                                         
#> 0.113200493 0.100849819 0.017751387 0.014557761 0.036059236 0.189705703 
#>                                                                         
#> 0.066567987 0.094952902 0.134558346 0.278559095 0.026516870 0.111742554 
#>                                                                         
#> 0.084155454 0.157989892 0.177766911 0.114571188 0.049226238 0.203173136 
#>                                                                         
#> 0.091849293 0.205547919 0.060845958 0.158015815 0.046359293 0.232575386 
#>                                                                         
#> 0.137319802 0.136222541 0.079276614 0.086949362 0.097816464 0.142949231 
#>                                                                         
#> 0.189355647 0.068946768 0.157777819 0.147841861 0.089935546 0.081470524 
#>                                                                         
#> 0.220062994 0.098349500 0.109887603 0.089793390 0.140441014 0.098749931 
#>                                                                         
#> 0.076907453 0.051086421 0.118174179 0.143909596 0.021730308 0.071987734 
#>                                                                         
#> 0.060500690 0.093426591 0.013214294 0.064814395 0.043914328 0.083214205 
#>                                                                         
#> 0.084326136 0.218887789 0.032937279 0.058994106 0.109038754 0.191452749 
#>                                                                         
#> 0.018462969 0.108441162 0.152939528 0.154946180 0.118692751 0.123944142 
#>                                                                         
#> 0.236430621 0.096582612 0.120173165 0.114496350 0.030300039 0.034273134 
#>                                                                         
#> 0.103515931 0.074379901 0.215112669 0.082963573 0.126233492 0.033984369 
#>                                                                         
#> 0.121708054 0.076332607 0.052172761 0.154099339 0.221710925 0.234573669 
#>                                                                         
#> 0.062546204 0.172850892 0.197754949 0.517010056 0.085366579 0.013132319 
#>                                                                         
#> 0.169668074 0.040556987 0.018367394 0.072151448 0.138334948 0.067696467 
#>                                                                         
#> 0.068514917 0.194995359 0.072805048 0.022559811 0.151493808 0.200041444 
#>                                                                         
#> 0.093796556 0.031855447 0.059559447 0.121684651 0.130274731 0.170722314 
#>                                                                         
#> 0.106286557 0.060903626 0.069649843 0.061988476 0.086173274 0.220119516 
#>                                                                         
#> 0.070478743 0.153159038 0.051209131 0.216778739 0.092717080 0.129694981 
#>                                                                         
#> 0.035830859 0.089244528 0.055779905 0.127375899 0.064523887 0.112197420 
#>                                                                         
#> 0.156284058 0.232539484 0.157287413 0.260908905 0.140369177 0.139137487 
#>                                                                         
#> 0.070621403 0.045212183 0.116939034 0.096154791 0.128796198 0.024364203 
#>                                                                         
#> 0.089163056 0.153916985 0.086710378 0.131983539 0.249787229 0.028013265 
#>                                                                         
#> 0.079058273 0.554779332 0.127971977 0.090635927 0.111161198 0.103846862 
#>                                                                         
#> 0.093454398 0.070106591 0.080744007 0.189754071 0.160341607 0.109949407 
#>                                                                         
#> 0.197406113 0.123662957 0.060785666 0.260062633 0.109927798 0.216809952 
#>                                                                         
#> 0.164524787 0.171088503 0.160957487 0.072486815 0.201629454 0.112729937 
#>                                                                         
#> 0.132071169 0.234807194 0.116100259 0.120705508 0.081137960 0.123686094 
#>                                                                         
#> 0.229760251 0.037105346 0.099300081 0.088750314 0.044466599 0.178029436 
#>                                                                         
#> 0.097793919 0.094263828 0.127639251 0.093077117 0.142998668 0.206131547 
#>                                                                         
#> 0.258765974 0.116915825 0.071118515 0.086882932 0.068562517 0.148090317 
#>                                                                         
#> 0.149061741 0.076189260 0.047721982 0.232792444 0.082887065 0.062115299 
#>                                                                         
#> 0.189552682 0.152432671 0.150549834 0.182692372 0.135297019 0.106317655 
#>                                                                         
#> 0.013367150 0.088646757 0.100492589 0.163009431 0.134116635 0.072714251 
#>                                                                         
#> 0.057485691 0.082203504 0.157803078 0.241762770 0.253885384 0.168157952 
#>                                                                         
#> 0.080248239 0.267702591 0.465191997 0.203559805 0.190115534 0.159100646 
#>                                                                         
#> 0.053856688 0.104172165 0.073687076 0.382998724 0.073418037 0.127450045 
#>                                                                         
#> 0.037581835 0.098329231 0.090228766 0.123051069 0.112338055 0.166450503 
#>                                                                         
#> 0.096390260 0.109356580 0.041403612 0.193333350 0.156218660 0.098250762 
#>                                                                         
#> 0.218700126 0.071202848 0.106136381 0.089202989 0.190453986 0.121803475 
#>                                                                         
#> 0.139278982 0.071639135 0.162749433 0.144151296 0.070511126 0.087914017 
#>                                                                         
#> 0.064852837 0.102119062 0.125879902 0.195457098 0.116728070 0.146614241 
#>                                                                         
#> 0.300096467 0.134538793 0.395929906 0.129359686 0.053940928 0.127103070 
#>                                                                         
#> 0.100472262 0.235683888 0.008928713 0.108986766 0.251207747 0.109679906 
#>                                                                         
#> 0.313698863 0.188250289 0.061122991 0.030870772 0.039447262 0.207780019 
#>                                                 
#> 0.162748397 0.129344936 0.272536722 0.787302658 
#> 
#> $alpha
#> [1] 0.28463
#> 
#> attr(,"class")
#> [1] "scree"