Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1
#>    46   92 0.93484418 0.09263485 0.98124334 0.07339669  1.007865763
#>    92   83 0.98124334 0.07339669 0.98271853 0.43035322  1.578026386
#>    83   54 0.98271853 0.43035322 0.93028285 0.33591704  0.960100374
#>    54   65 0.93028285 0.33591704 0.80340109 0.26420285  0.912489246
#>    65   46 0.80340109 0.26420285 0.93484418 0.09263485  0.914828793
#>    68   13 0.16495327 0.73525392 0.14807511 0.54981046  0.196518181
#>    13   85 0.14807511 0.54981046 0.17916117 0.42305358  0.253715021
#>    85   11 0.17916117 0.42305358 0.48092373 0.46233593  0.327268194
#>    11   60 0.48092373 0.46233593 0.43559205 0.62550755  0.373220099
#>    60   38 0.43559205 0.62550755 0.34724956 0.66089788  0.368041365
#>    38    5 0.34724956 0.66089788 0.26452891 0.76738317  0.274523177
#>     5   68 0.26452891 0.76738317 0.16495327 0.73525392  0.236997598
#>    74   49 0.56490336 0.36806005 0.70250012 0.44553169  0.613104568
#>    49  100 0.70250012 0.44553169 0.71229484 0.53595724  0.673523343
#>   100   48 0.71229484 0.53595724 0.66752959 0.61408579  0.669494086
#>    48    9 0.66752959 0.61408579 0.56290849 0.59957842  0.618038357
#>     9   12 0.56290849 0.59957842 0.59136671 0.53050032  0.609964648
#>    12   74 0.59136671 0.53050032 0.56490336 0.36806005  0.612241016
#>    56   47 0.54730000 0.00000000 0.94514884 0.02020321  0.794830698
#>    47   19 0.94514884 0.02020321 0.85792369 0.09050501  0.795091061
#>    19   32 0.85792369 0.09050501 0.71782576 0.09966757  0.779691647
#>    32   91 0.71782576 0.09966757 0.63124334 0.09343351  0.676864560
#>    91   51 0.63124334 0.09343351 0.60731071 0.04416577  0.642255777
#>    51   56 0.60731071 0.04416577 0.54730000 0.00000000  0.600163896
#>    41   88 0.31111638 0.99178373 0.17183037 0.85015500  0.282413511
#>    88   68 0.17183037 0.85015500 0.16495327 0.73525392  0.202037198
#>    38   78 0.34724956 0.66089788 0.38743682 0.78117396  0.330578604
#>    78   33 0.38743682 0.78117396 0.39924336 0.93342096  0.315994798
#>    33   41 0.39924336 0.93342096 0.31111638 0.99178373  0.298703899
#>    69   23 0.37633909 0.20939957 0.29584690 0.19498136  0.350874660
#>    23   31 0.29584690 0.19498136 0.23422817 0.08900089  0.334066777
#>    31   17 0.23422817 0.08900089 0.25044226 0.01965548  0.331340789
#>    17   56 0.25044226 0.01965548 0.54730000 0.00000000  0.389314603
#>    56   44 0.54730000 0.00000000 0.51190272 0.04310090  0.435358433
#>    44   79 0.51190272 0.04310090 0.49350537 0.21158316  0.492643525
#>    79   69 0.49350537 0.21158316 0.37633909 0.20939957  0.436016408
#>    76    8 0.11267169 0.00293590 0.09904676 0.04818352  0.061373266
#>     8   63 0.09904676 0.04818352 0.00000000 0.11657167  0.054448963
#>    76   15 0.11267169 0.00293590 0.19694507 0.10807704  0.168599954
#>    15   90 0.19694507 0.10807704 0.06042147 0.20950723  0.112211477
#>    90   63 0.06042147 0.20950723 0.00000000 0.11657167  0.081978751
#>    48   52 0.66752959 0.61408579 0.60151501 0.66336625  0.614513186
#>    52   29 0.60151501 0.66336625 0.53264745 0.72159273  0.538082318
#>    29   60 0.53264745 0.72159273 0.43559205 0.62550755  0.506973544
#>    11   28 0.48092373 0.46233593 0.58198510 0.53274422  0.498376236
#>    28   12 0.58198510 0.53274422 0.59136671 0.53050032  0.596584870
#>    87  100 0.79445774 0.48704020 0.71229484 0.53595724  0.760311475
#>    49   96 0.70250012 0.44553169 0.74428326 0.36340351  0.750870181
#>    96   35 0.74428326 0.36340351 0.76863026 0.36289012  0.757608787
#>    35   87 0.76863026 0.36289012 0.79445774 0.48704020  0.794566109
#>    74   28 0.56490336 0.36806005 0.58198510 0.53274422  0.563582211
#>    85   73 0.17916117 0.42305358 0.39583621 0.32211050  0.327884645
#>    73   93 0.39583621 0.32211050 0.44817701 0.34783890  0.374944949
#>    93   74 0.44817701 0.34783890 0.56490336 0.36806005  0.500137422
#>    68   75 0.16495327 0.73525392 0.09402421 0.59634630  0.179078664
#>    75   99 0.09402421 0.59634630 0.11583442 0.55564475  0.139519518
#>    99   85 0.11583442 0.55564475 0.17916117 0.42305358  0.117930300
#>    64   43 0.86149855 1.00000000 0.53938992 0.98484200  0.662479423
#>    43   98 0.53938992 0.98484200 0.54826359 0.97184393  0.642725272
#>    98   25 0.54826359 0.97184393 0.58468734 0.90608534  0.645219933
#>    25   30 0.58468734 0.90608534 0.79196744 0.87661393  0.688158722
#>    30   64 0.79196744 0.87661393 0.86149855 1.00000000  0.809288955
#>    76   17 0.11267169 0.00293590 0.25044226 0.01965548  0.178495749
#>    31   53 0.23422817 0.08900089 0.20275229 0.19794476  0.242061843
#>    53   90 0.20275229 0.19794476 0.06042147 0.20950723  0.127844936
#>    96   18 0.74428326 0.36340351 0.73540386 0.31173371  0.724287326
#>    18   65 0.73540386 0.31173371 0.80340109 0.26420285  0.779102916
#>    65   35 0.80340109 0.26420285 0.76863026 0.36289012  0.810857062
#>    76   56 0.11267169 0.00293590 0.54730000 0.00000000  0.322222585
#>    96   74 0.74428326 0.36340351 0.56490336 0.36806005  0.654244083
#>    74   58 0.56490336 0.36806005 0.69904635 0.24731571  0.643883924
#>    58   65 0.69904635 0.24731571 0.80340109 0.26420285  0.750408669
#>    47   46 0.94514884 0.02020321 0.93484418 0.09263485  0.897524529
#>    46   27 0.93484418 0.09263485 0.79179577 0.20728057  0.893717069
#>    27   32 0.79179577 0.20728057 0.71782576 0.09966757  0.790106633
#>    83   26 0.98271853 0.43035322 0.98559755 0.52443893  1.017095507
#>    26   36 0.98559755 0.52443893 0.86377218 0.61395502  0.923185480
#>    36   97 0.86377218 0.61395502 0.87244321 0.45782708  0.892258716
#>    97   45 0.87244321 0.45782708 0.92737008 0.44606594  0.910470360
#>    45   83 0.92737008 0.44606594 0.98271853 0.43035322  0.962898569
#>    64   42 0.86149855 1.00000000 0.84607948 0.87761327  0.854046483
#>    42   57 0.84607948 0.87761327 0.97506537 0.85264403  0.912478840
#>    93   84 0.44817701 0.34783890 0.44944793 0.32226553  0.486253495
#>    84   79 0.44944793 0.32226553 0.49350537 0.21158316  0.458354021
#>    79   67 0.49350537 0.21158316 0.66394106 0.17812350  0.588699080
#>    67   58 0.66394106 0.17812350 0.69904635 0.24731571  0.596569103
#>    69   59 0.37633909 0.20939957 0.35972935 0.23441941  0.336330816
#>    59   82 0.35972935 0.23441941 0.16829109 0.39851301  0.264563882
#>    82   53 0.16829109 0.39851301 0.20275229 0.19794476  0.252907572
#>    25   29 0.58468734 0.90608534 0.53264745 0.72159273  0.634567717
#>    52   80 0.60151501 0.66336625 0.69532481 0.65606066  0.653346071
#>    80   77 0.69532481 0.65606066 0.86733268 0.62820944  0.781612035
#>    77   34 0.86733268 0.62820944 0.84663786 0.75339346  0.806489135
#>    34   66 0.84663786 0.75339346 0.77403772 0.77116999  0.797473238
#>    66   25 0.77403772 0.77116999 0.58468734 0.90608534  0.661004542
#>    98   95 0.54826359 0.97184393 0.44400132 0.98144025  0.488699508
#>    95   33 0.44400132 0.98144025 0.39924336 0.93342096  0.488351052
#>    33   81 0.39924336 0.93342096 0.42833609 0.74980058  0.482040662
#>    81   29 0.42833609 0.74980058 0.53264745 0.72159273  0.505665831
#>    66   30 0.77403772 0.77116999 0.79196744 0.87661393  0.681197449
#>    83   61 0.98271853 0.43035322 1.00000000 0.68184962  3.062107187
#>    61   70 1.00000000 0.68184962 0.98618786 0.61463967  1.388908431
#>    70   36 0.98618786 0.61463967 0.86377218 0.61395502  0.925228124
#>    67   32 0.66394106 0.17812350 0.71782576 0.09966757  0.733468869
#>    81   60 0.42833609 0.74980058 0.43559205 0.62550755  0.468070095
#>    66   42 0.77403772 0.77116999 0.84607948 0.87761327  0.820122918
#>    67   91 0.66394106 0.17812350 0.63124334 0.09343351  0.672399695
#>    95   41 0.44400132 0.98144025 0.31111638 0.99178373  0.378421698
#>    78   81 0.38743682 0.78117396 0.42833609 0.74980058  0.473542101
#>    77   55 0.86733268 0.62820944 0.97562598 0.73950744  0.906013657
#>    55   57 0.97562598 0.73950744 0.97506537 0.85264403  0.916442830
#>    54   37 0.93028285 0.33591704 0.89499171 0.43437392  0.851780602
#>    37   97 0.89499171 0.43437392 0.87244321 0.45782708  0.832703497
#>    97   87 0.87244321 0.45782708 0.79445774 0.48704020  0.813201862
#>    80  100 0.69532481 0.65606066 0.71229484 0.53595724  0.772216227
#>    70   77 0.98618786 0.61463967 0.86733268 0.62820944  0.925025155
#>    54   45 0.93028285 0.33591704 0.92737008 0.44606594  0.928956334
#>    85   94 0.17916117 0.42305358 0.06212154 0.53334174  0.094220887
#>    94   62 0.06212154 0.53334174 0.05175016 0.21150235 -0.103127092
#>    62   72 0.05175016 0.21150235 0.12911644 0.31037147  0.006701703
#>    72   40 0.12911644 0.31037147 0.12795485 0.35006649  0.009551780
#>    40   82 0.12795485 0.35006649 0.16829109 0.39851301  0.157425188
#>    82   85 0.16829109 0.39851301 0.17916117 0.42305358  0.102073586
#>    72   82 0.12911644 0.31037147 0.16829109 0.39851301  0.198592640
#>    50   88 0.11125837 0.88319039 0.17183037 0.85015500  0.195705637
#>    95   43 0.44400132 0.98144025 0.53938992 0.98484200  0.455748612
#>    56   86 0.54730000 0.00000000 0.58558065 0.07249554  0.554910515
#>    86   79 0.58558065 0.07249554 0.49350537 0.21158316  0.520232769
#>    47   92 0.94514884 0.02020321 0.98124334 0.07339669  0.947456162
#>    89   75 0.06151379 0.55998002 0.09402421 0.59634630 -0.117819898
#>    91   86 0.63124334 0.09343351 0.58558065 0.07249554  0.613654794
#>    72   90 0.12911644 0.31037147 0.06042147 0.20950723  0.133983806
#>    61   57 1.00000000 0.68184962 0.97506537 0.85264403  1.192367689
#>    77   61 0.86733268 0.62820944 1.00000000 0.68184962  0.922354507
#>    69   84 0.37633909 0.20939957 0.44944793 0.32226553  0.423676146
#>    84   73 0.44944793 0.32226553 0.39583621 0.32211050  0.422822718
#>    73   82 0.39583621 0.32211050 0.16829109 0.39851301  0.269487110
#>    94   63 0.06212154 0.53334174 0.00000000 0.11657167 -0.382355602
#>    94   89 0.06212154 0.53334174 0.06151379 0.55998002 -1.254813153
#>    79   91 0.49350537 0.21158316 0.63124334 0.09343351  0.572795062
#>    94   99 0.06212154 0.53334174 0.11583442 0.55564475  0.092674114
#>     1   92 0.96068100 0.32697426 0.98124334 0.07339669  1.007865763
#>     1   83 0.96068100 0.32697426 0.98271853 0.43035322  1.578026386
#>     1   54 0.96068100 0.32697426 0.93028285 0.33591704  0.960100374
#>     1   65 0.96068100 0.32697426 0.80340109 0.26420285  0.912489246
#>     1   46 0.96068100 0.32697426 0.93484418 0.09263485  0.914828793
#>     2   13 0.29487108 0.61421693 0.14807511 0.54981046  0.196518181
#>     2   85 0.29487108 0.61421693 0.17916117 0.42305358  0.253715021
#>     2   11 0.29487108 0.61421693 0.48092373 0.46233593  0.327268194
#>     2   60 0.29487108 0.61421693 0.43559205 0.62550755  0.373220099
#>     2   38 0.29487108 0.61421693 0.34724956 0.66089788  0.368041365
#>     2    5 0.29487108 0.61421693 0.26452891 0.76738317  0.274523177
#>     2   68 0.29487108 0.61421693 0.16495327 0.73525392  0.236997598
#>     3   49 0.63017433 0.53115547 0.70250012 0.44553169  0.613104568
#>     3  100 0.63017433 0.53115547 0.71229484 0.53595724  0.673523343
#>     3   48 0.63017433 0.53115547 0.66752959 0.61408579  0.669494086
#>     3    9 0.63017433 0.53115547 0.56290849 0.59957842  0.618038357
#>     3   12 0.63017433 0.53115547 0.59136671 0.53050032  0.609964648
#>     3   74 0.63017433 0.53115547 0.56490336 0.36806005  0.612241016
#>     4   47 0.64509129 0.02029297 0.94514884 0.02020321  0.794830698
#>     4   19 0.64509129 0.02029297 0.85792369 0.09050501  0.795091061
#>     4   32 0.64509129 0.02029297 0.71782576 0.09966757  0.779691647
#>     4   91 0.64509129 0.02029297 0.63124334 0.09343351  0.676864560
#>     4   51 0.64509129 0.02029297 0.60731071 0.04416577  0.642255777
#>     4   56 0.64509129 0.02029297 0.54730000 0.00000000  0.600163896
#>     5   88 0.26452891 0.76738317 0.17183037 0.85015500  0.282413511
#>     5   78 0.26452891 0.76738317 0.38743682 0.78117396  0.330578604
#>     5   33 0.26452891 0.76738317 0.39924336 0.93342096  0.315994798
#>     5   41 0.26452891 0.76738317 0.31111638 0.99178373  0.298703899
#>     6   23 0.42925181 0.06910247 0.29584690 0.19498136  0.350874660
#>     6   31 0.42925181 0.06910247 0.23422817 0.08900089  0.334066777
#>     6   17 0.42925181 0.06910247 0.25044226 0.01965548  0.331340789
#>     6   56 0.42925181 0.06910247 0.54730000 0.00000000  0.389314603
#>     6   44 0.42925181 0.06910247 0.51190272 0.04310090  0.435358433
#>     6   79 0.42925181 0.06910247 0.49350537 0.21158316  0.492643525
#>     6   69 0.42925181 0.06910247 0.37633909 0.20939957  0.436016408
#>     7    8 0.01790102 0.04091916 0.09904676 0.04818352  0.061373266
#>     7   63 0.01790102 0.04091916 0.00000000 0.11657167  0.054448963
#>     8   15 0.09904676 0.04818352 0.19694507 0.10807704  0.168599954
#>     8   90 0.09904676 0.04818352 0.06042147 0.20950723  0.112211477
#>     9   52 0.56290849 0.59957842 0.60151501 0.66336625  0.614513186
#>     9   29 0.56290849 0.59957842 0.53264745 0.72159273  0.538082318
#>     9   60 0.56290849 0.59957842 0.43559205 0.62550755  0.506973544
#>     9   11 0.56290849 0.59957842 0.48092373 0.46233593  0.486895232
#>     9   28 0.56290849 0.59957842 0.58198510 0.53274422  0.498376236
#>    10  100 0.75520792 0.47371409 0.71229484 0.53595724  0.760311475
#>    10   49 0.75520792 0.47371409 0.70250012 0.44553169  0.712509667
#>    10   96 0.75520792 0.47371409 0.74428326 0.36340351  0.750870181
#>    10   35 0.75520792 0.47371409 0.76863026 0.36289012  0.757608787
#>    10   87 0.75520792 0.47371409 0.79445774 0.48704020  0.794566109
#>    11   73 0.48092373 0.46233593 0.39583621 0.32211050  0.327884645
#>    11   93 0.48092373 0.46233593 0.44817701 0.34783890  0.374944949
#>    11   74 0.48092373 0.46233593 0.56490336 0.36806005  0.500137422
#>    13   75 0.14807511 0.54981046 0.09402421 0.59634630  0.179078664
#>    13   99 0.14807511 0.54981046 0.11583442 0.55564475  0.139519518
#>    14   43 0.74272066 0.97951007 0.53938992 0.98484200  0.662479423
#>    14   98 0.74272066 0.97951007 0.54826359 0.97184393  0.642725272
#>    14   25 0.74272066 0.97951007 0.58468734 0.90608534  0.645219933
#>    14   30 0.74272066 0.97951007 0.79196744 0.87661393  0.688158722
#>    14   64 0.74272066 0.97951007 0.86149855 1.00000000  0.809288955
#>    15   17 0.19694507 0.10807704 0.25044226 0.01965548  0.178495749
#>    15   31 0.19694507 0.10807704 0.23422817 0.08900089  0.186257051
#>    15   53 0.19694507 0.10807704 0.20275229 0.19794476  0.242061843
#>    16   18 0.75487813 0.33953571 0.73540386 0.31173371  0.724287326
#>    16   65 0.75487813 0.33953571 0.80340109 0.26420285  0.779102916
#>    16   35 0.75487813 0.33953571 0.76863026 0.36289012  0.810857062
#>    16   96 0.75487813 0.33953571 0.74428326 0.36340351  0.756273174
#>    18   74 0.73540386 0.31173371 0.56490336 0.36806005  0.654244083
#>    18   58 0.73540386 0.31173371 0.69904635 0.24731571  0.643883924
#>    19   46 0.85792369 0.09050501 0.93484418 0.09263485  0.897524529
#>    19   27 0.85792369 0.09050501 0.79179577 0.20728057  0.893717069
#>    20   26 0.96800118 0.50623348 0.98559755 0.52443893  1.017095507
#>    20   36 0.96800118 0.50623348 0.86377218 0.61395502  0.923185480
#>    20   97 0.96800118 0.50623348 0.87244321 0.45782708  0.892258716
#>    20   45 0.96800118 0.50623348 0.92737008 0.44606594  0.910470360
#>    20   83 0.96800118 0.50623348 0.98271853 0.43035322  0.962898569
#>    21   42 0.91567025 0.94135180 0.84607948 0.87761327  0.854046483
#>    21   57 0.91567025 0.94135180 0.97506537 0.85264403  0.912478840
#>    22   84 0.50460303 0.30180607 0.44944793 0.32226553  0.486253495
#>    22   79 0.50460303 0.30180607 0.49350537 0.21158316  0.458354021
#>    22   67 0.50460303 0.30180607 0.66394106 0.17812350  0.588699080
#>    22   58 0.50460303 0.30180607 0.69904635 0.24731571  0.596569103
#>    22   74 0.50460303 0.30180607 0.56490336 0.36806005  0.601674592
#>    22   93 0.50460303 0.30180607 0.44817701 0.34783890  0.505992846
#>    23   59 0.29584690 0.19498136 0.35972935 0.23441941  0.336330816
#>    23   82 0.29584690 0.19498136 0.16829109 0.39851301  0.264563882
#>    23   53 0.29584690 0.19498136 0.20275229 0.19794476  0.252907572
#>    24   29 0.73198053 0.71551183 0.53264745 0.72159273  0.634567717
#>    24   52 0.73198053 0.71551183 0.60151501 0.66336625  0.633930626
#>    24   80 0.73198053 0.71551183 0.69532481 0.65606066  0.653346071
#>    24   77 0.73198053 0.71551183 0.86733268 0.62820944  0.781612035
#>    24   34 0.73198053 0.71551183 0.84663786 0.75339346  0.806489135
#>    24   66 0.73198053 0.71551183 0.77403772 0.77116999  0.797473238
#>    24   25 0.73198053 0.71551183 0.58468734 0.90608534  0.661004542
#>    25   95 0.58468734 0.90608534 0.44400132 0.98144025  0.488699508
#>    25   33 0.58468734 0.90608534 0.39924336 0.93342096  0.488351052
#>    25   81 0.58468734 0.90608534 0.42833609 0.74980058  0.482040662
#>    26   61 0.98559755 0.52443893 1.00000000 0.68184962  3.062107187
#>    26   70 0.98559755 0.52443893 0.98618786 0.61463967  1.388908431
#>    27   67 0.79179577 0.20728057 0.66394106 0.17812350  0.728773872
#>    27   65 0.79179577 0.20728057 0.80340109 0.26420285  0.913184640
#>    27   58 0.79179577 0.20728057 0.69904635 0.24731571  0.752991339
#>    30   42 0.79196744 0.87661393 0.84607948 0.87761327  0.820122918
#>    34   55 0.84663786 0.75339346 0.97562598 0.73950744  0.906013657
#>    34   57 0.84663786 0.75339346 0.97506537 0.85264403  0.916442830
#>    34   42 0.84663786 0.75339346 0.84607948 0.87761327  0.901013463
#>    35   54 0.76863026 0.36289012 0.93028285 0.33591704  0.847062423
#>    35   37 0.76863026 0.36289012 0.89499171 0.43437392  0.851780602
#>    35   97 0.76863026 0.36289012 0.87244321 0.45782708  0.832703497
#>    36  100 0.86377218 0.61395502 0.71229484 0.53595724  0.772216227
#>    36   87 0.86377218 0.61395502 0.79445774 0.48704020  0.789481949
#>    36   77 0.86377218 0.61395502 0.86733268 0.62820944  0.925025155
#>    36   80 0.86377218 0.61395502 0.69532481 0.65606066  0.781326371
#>    37   45 0.89499171 0.43437392 0.92737008 0.44606594  0.928956334
#>    38   81 0.34724956 0.66089788 0.42833609 0.74980058  0.408683736
#>    39   94 0.12366515 0.36600142 0.06212154 0.53334174  0.094220887
#>    39   62 0.12366515 0.36600142 0.05175016 0.21150235 -0.103127092
#>    39   72 0.12366515 0.36600142 0.12911644 0.31037147  0.006701703
#>    39   40 0.12366515 0.36600142 0.12795485 0.35006649  0.009551780
#>    39   82 0.12366515 0.36600142 0.16829109 0.39851301  0.157425188
#>    39   85 0.12366515 0.36600142 0.17916117 0.42305358  0.102073586
#>    41   43 0.31111638 0.99178373 0.53938992 0.98484200  0.455748612
#>    41   64 0.31111638 0.99178373 0.86149855 1.00000000  0.533353533
#>    44   86 0.51190272 0.04310090 0.58558065 0.07249554  0.554910515
#>    48   80 0.66752959 0.61408579 0.69532481 0.65606066  0.733981973
#>    50   75 0.11125837 0.88319039 0.09402421 0.59634630 -0.117819898
#>    50   68 0.11125837 0.88319039 0.16495327 0.73525392 -0.031108428
#>    51   86 0.60731071 0.04416577 0.58558065 0.07249554  0.613654794
#>    53   72 0.20275229 0.19794476 0.12911644 0.31037147  0.250184254
#>    55   61 0.97562598 0.73950744 1.00000000 0.68184962  0.922354507
#>    59   84 0.35972935 0.23441941 0.44944793 0.32226553  0.423676146
#>    59   73 0.35972935 0.23441941 0.39583621 0.32211050  0.422822718
#>    61   92 1.00000000 0.68184962 0.98124334 0.07339669  5.702668727
#>    62   63 0.05175016 0.21150235 0.00000000 0.11657167 -0.382355602
#>    62   90 0.05175016 0.21150235 0.06042147 0.20950723  0.043219035
#>    63   89 0.00000000 0.11657167 0.06151379 0.55998002 -1.254813153
#>    71   99 0.07142574 0.55766050 0.11583442 0.55564475  0.092674114
#>    71   75 0.07142574 0.55766050 0.09402421 0.59634630  0.094248276
#>    71   89 0.07142574 0.55766050 0.06151379 0.55998002  0.072167994
#>    71   94 0.07142574 0.55766050 0.06212154 0.53334174  0.063634009
#>     7   76 0.01790102 0.04091916 0.11267169 0.00293590  0.061373266
#>    21   64 0.91567025 0.94135180 0.86149855 1.00000000  0.854046483
#>    41   50 0.31111638 0.99178373 0.11125837 0.88319039  0.195705637
#>    50   89 0.11125837 0.88319039 0.06151379 0.55998002 -0.117819898
#>           my1           mx2          my2 bp1 bp2
#>   0.203177946   0.947456162  0.057480295   0   0
#>   0.249411678   5.702668727  0.232365769   0   0
#>   0.381136409   0.941736608  0.391332889   0   0
#>   0.219297594   0.847062423  0.335055296   0   0
#>   0.213435629   0.913184640  0.212175996   0   0
#>   0.638891221   0.179078664  0.640478480   0   0
#>   0.508527525   0.117930300  0.475227462   0   0
#>   0.464006279   0.327884645  0.459270769   0   0
#>   0.520296884   0.486895232  0.551877650   0   0
#>   0.584842196   0.408683736  0.686295010   0   0
#>   0.689774528   0.330578604  0.733319893   0   0
#>   0.682340727   0.202037198  0.790690703   0   0
#>   0.443378339   0.654691546  0.369516041   0   0
#>   0.494413644   0.712509667  0.490190724   0   0
#>   0.563322557   0.733981973  0.600272127   0   0
#>   0.586500359   0.614513186  0.611922432   0   0
#>   0.578563186   0.596584870  0.573051095   0   0
#>   0.443723934   0.567399597  0.451029107   0   0
#>  -0.947070467   0.922141644 -3.454122929   0   1
#>  -0.076715049   0.897524529  0.050376649   0   0
#>  -0.030035099   0.790106633  0.129212727   0   0
#>   0.064189925   0.672399695  0.126200643   0   0
#>   0.057637322   0.613654794  0.071530730   0   0
#>  -0.008976393   0.567160292  0.035867597   0   0
#>   0.880706436   0.195705637  0.965980047   0   0
#>   0.790690703   0.103326725  0.796598761   0   0
#>   0.733319893   0.372473417  0.719321781   0   0
#>   0.863295475   0.473542101  0.851077913   0   0
#>   0.877324413   0.378421698  0.997697196   0   0
#>   0.119669218   0.336330816  0.200862777   0   0
#>   0.101856431   0.247883123  0.151964976   0   0
#>   0.075139139   0.186257051  0.041216199   0   0
#>  -0.134505013   0.322222585 -1.147799316   0   0
#>  -0.055848034   0.554910515  0.042335987   0   0
#>   0.126243474   0.520232769  0.129256070   0   0
#>   0.151780225   0.434147428  0.252065341   0   0
#>   0.012164124   0.168599954  0.044452161   0   0
#>   0.089511332   0.081978751  0.129382805   0   0
#>   0.044452161   0.178495749  0.036520423   0   0
#>   0.136621342   0.127844936  0.157663756   0   0
#>   0.129382805   0.043219035  0.154582187   0   0
#>   0.611922432   0.648213789  0.657066715   0   0
#>   0.658180936   0.633930626  0.771545833   0   0
#>   0.650465580   0.468070095  0.689761858   0   0
#>   0.545019225   0.563582211  0.451425063   0   0
#>   0.573051095   0.567399597  0.451029107   0   0
#>   0.523147327   0.789481949  0.572143154   0   0
#>   0.418447425   0.654691546  0.369516041   0   0
#>   0.417780065   0.756273174  0.354440386   0   0
#>   0.422256116   0.813201862  0.418379240   0   0
#>   0.451425063   0.567399597  0.451029107   0   0
#>   0.459270769   0.282789764  0.362474288   0   0
#>   0.430715006   0.422608635  0.333749967   0   0
#>   0.394909339   0.505992846  0.361108972   0   0
#>   0.640478480  -0.031108428  0.747804289   0   0
#>   0.594530933   0.094248276  0.570272014   0   0
#>   0.475227462   0.117770638  0.475151206   0   0
#>   1.799176164   0.533353533  4.543111056   0   0
#>   1.045860104   0.493101694  0.943713420   0   0
#>   0.982581276   0.488699508  0.895884461   0   0
#>   0.890163379   0.681197449  0.841202932   0   0
#>   0.948137116   0.817800381  0.943340716   0   0
#>   0.036520423   0.322222585 -1.147799316   0   0
#>   0.150283101   0.247883123  0.151964976   0   0
#>   0.157663756   0.133983806  0.233231537   0   0
#>   0.340241934   0.654244083  0.352278790   0   0
#>   0.301845643   0.750408669  0.260795910   0   0
#>   0.322298936   0.847062423  0.335055296   0   0
#>  -1.147799316   0.306681352 -3.448510862   0   1
#>   0.352278790   0.654691546  0.369516041   0   0
#>   0.320918461   0.601674592  0.274025281   0   0
#>   0.260795910   0.752991339  0.244836201   0   0
#>   0.050376649   0.947456162  0.057480295   0   0
#>   0.187885470   0.913184640  0.212175996   0   0
#>   0.129212727   0.733468869  0.168143863   0   0
#>   0.476388188   3.062107187  0.413810847   0   0
#>   0.567156409   0.925228124  0.569936310   0   0
#>   0.537232352   0.856990130  0.535273605   0   0
#>   0.501281159   0.902527109  0.464184598   0   0
#>   0.465876423   0.941736608  0.391332889   0   0
#>   0.938774197   0.817800381  0.943340716   0   0
#>   0.874976797   0.901013463  0.815749043   0   0
#>   0.336912912   0.422608635  0.333749967   0   0
#>   0.261700847   0.434147428  0.252065341   0   0
#>   0.245668055   0.572795062  0.164656717   0   0
#>   0.255806864   0.728773872  0.188731581   0   0
#>   0.200862777   0.423676146  0.258848106   0   0
#>   0.317112134   0.269487110  0.322855773   0   0
#>   0.309806980   0.250184254  0.309339065   0   0
#>   0.792429809   0.505665831  0.828789202   0   0
#>   0.722969396   0.648213789  0.657066715   0   0
#>   0.643884638   0.781326371  0.642120391   0   0
#>   0.682453674   0.906013657  0.698906591   0   0
#>   0.709742323   0.823344057  0.815399912   0   0
#>   0.812862683   0.681197449  0.841202932   0   0
#>   0.895884461   0.493101694  0.943713420   0   0
#>   0.895233901   0.378421698  0.997697196   0   0
#>   0.852424422   0.473542101  0.851077913   0   0
#>   0.828789202   0.468070095  0.689761858   0   0
#>   0.841202932   0.820122918  0.817580007   0   0
#>   0.413810847   5.702668727  0.232365769   0   0
#>   0.566901827   0.931272838  0.660949340   0   0
#>   0.569936310   0.925025155  0.606227013   0   0
#>   0.168143863   0.672399695  0.126200643   0   0
#>   0.689761858   0.408683736  0.686295010   0   0
#>   0.817580007   0.823344057  0.815399912   0   0
#>   0.126200643   0.572795062  0.164656717   0   0
#>   0.997697196   0.455748612  1.991132772   0   0
#>   0.851077913   0.372473417  0.719321781   0   0
#>   0.698906591   0.922354507  0.683006893   0   0
#>   0.795783865   1.192367689  0.797151112   0   0
#>   0.363331856   0.928956334  0.390994929   0   0
#>   0.397054330   0.902527109  0.464184598   0   0
#>   0.418379240   0.856990130  0.535273605   0   0
#>   0.605674440   0.733981973  0.600272127   0   0
#>   0.606227013   0.931272838  0.660949340   0   0
#>   0.390994929   0.941736608  0.391332889   0   0
#>   0.450159817   0.117770638  0.475151206   0   0
#>   0.377580123  -0.382355602  0.386578355   0   0
#>   0.326457890   0.071159428  0.276018942   0   0
#>   0.326737174   0.198592640  0.332269043   0   0
#>   0.366544783   0.198592640  0.332269043   0   0
#>   0.442521306   0.282789764  0.362474288   0   0
#>   0.332269043   0.250184254  0.309339065   0   0
#>   0.965980047   0.103326725  0.796598761   0   0
#>   1.991132772   0.493101694  0.943713420   0   0
#>   0.042335987   0.567160292  0.035867597   0   0
#>   0.129256070   0.571594812  0.163257474   0   0
#>   0.057480295  13.819881783 -8.677118931   0   1
#>   0.753014085   0.072167994  0.583170300   0   0
#>   0.071530730   0.571594812  0.163257474   0   0
#>   0.233231537   0.071159428  0.276018942   0   0
#>   0.797151112   6.925173003  1.634095246   0   1
#>   0.683006893   0.931272838  0.660949340   0   0
#>   0.258848106   0.434147428  0.252065341   0   0
#>   0.259719724   0.422608635  0.333749967   0   0
#>   0.322855773   0.282789764  0.362474288   0   0
#>   0.386578355  -1.254813153  0.516622253   0   0
#>   0.516622253   0.063634009  0.546702318   0   0
#>   0.164656717   0.571594812  0.163257474   0   0
#>   0.535591758   0.117770638  0.475151206   0   0
#>   0.203177946   1.578026386  0.249411678   0   0
#>   0.249411678   0.960100374  0.381136409   0   0
#>   0.381136409   0.912489246  0.219297594   0   0
#>   0.219297594   0.914828793  0.213435629   0   0
#>   0.213435629   1.007865763  0.203177946   0   0
#>   0.638891221   0.253715021  0.508527525   0   0
#>   0.508527525   0.327268194  0.464006279   0   0
#>   0.464006279   0.373220099  0.520296884   0   0
#>   0.520296884   0.368041365  0.584842196   0   0
#>   0.584842196   0.274523177  0.689774528   0   0
#>   0.689774528   0.236997598  0.682340727   0   0
#>   0.682340727   0.196518181  0.638891221   0   0
#>   0.443378339   0.673523343  0.494413644   0   0
#>   0.494413644   0.669494086  0.563322557   0   0
#>   0.563322557   0.618038357  0.586500359   0   0
#>   0.586500359   0.609964648  0.578563186   0   0
#>   0.578563186   0.612241016  0.443723934   0   0
#>   0.443723934   0.613104568  0.443378339   0   0
#>  -0.947070467   0.795091061 -0.076715049   0   0
#>  -0.076715049   0.779691647 -0.030035099   0   0
#>  -0.030035099   0.676864560  0.064189925   0   0
#>   0.064189925   0.642255777  0.057637322   0   0
#>   0.057637322   0.600163896 -0.008976393   0   0
#>  -0.008976393   0.794830698 -0.947070467   0   0
#>   0.880706436   0.202037198  0.790690703   0   0
#>   0.733319893   0.315994798  0.863295475   0   0
#>   0.863295475   0.298703899  0.877324413   0   0
#>   0.877324413   0.282413511  0.880706436   0   0
#>   0.119669218   0.334066777  0.101856431   0   0
#>   0.101856431   0.331340789  0.075139139   0   0
#>   0.075139139   0.389314603 -0.134505013   0   0
#>  -0.134505013   0.435358433 -0.055848034   0   0
#>  -0.055848034   0.492643525  0.126243474   0   0
#>   0.126243474   0.436016408  0.151780225   0   0
#>   0.151780225   0.350874660  0.119669218   0   0
#>   0.012164124   0.054448963  0.089511332   0   0
#>   0.089511332 -12.463032754 -2.872395051   0   1
#>   0.044452161   0.112211477  0.136621342   0   0
#>   0.136621342   0.081978751  0.129382805   0   0
#>   0.611922432   0.538082318  0.658180936   0   0
#>   0.658180936   0.506973544  0.650465580   0   0
#>   0.650465580   0.486895232  0.551877650   0   0
#>   0.551877650   0.498376236  0.545019225   0   0
#>   0.545019225   0.596584870  0.573051095   0   0
#>   0.523147327   0.712509667  0.490190724   0   0
#>   0.490190724   0.750870181  0.418447425   0   0
#>   0.418447425   0.757608787  0.417780065   0   0
#>   0.417780065   0.794566109  0.422256116   0   0
#>   0.422256116   0.760311475  0.523147327   0   0
#>   0.459270769   0.374944949  0.430715006   0   0
#>   0.430715006   0.500137422  0.394909339   0   0
#>   0.394909339   0.563582211  0.451425063   0   0
#>   0.640478480   0.139519518  0.594530933   0   0
#>   0.594530933   0.117930300  0.475227462   0   0
#>   1.799176164   0.642725272  1.045860104   0   0
#>   1.045860104   0.645219933  0.982581276   0   0
#>   0.982581276   0.688158722  0.890163379   0   0
#>   0.890163379   0.809288955  0.948137116   0   0
#>   0.948137116   0.662479423  1.799176164   0   0
#>   0.036520423   0.186257051  0.041216199   0   0
#>   0.041216199   0.242061843  0.150283101   0   0
#>   0.150283101   0.127844936  0.157663756   0   0
#>   0.340241934   0.779102916  0.301845643   0   0
#>   0.301845643   0.810857062  0.322298936   0   0
#>   0.322298936   0.756273174  0.354440386   0   0
#>   0.354440386   0.724287326  0.340241934   0   0
#>   0.352278790   0.643883924  0.320918461   0   0
#>   0.320918461   0.750408669  0.260795910   0   0
#>   0.050376649   0.893717069  0.187885470   0   0
#>   0.187885470   0.790106633  0.129212727   0   0
#>   0.476388188   0.923185480  0.567156409   0   0
#>   0.567156409   0.892258716  0.537232352   0   0
#>   0.537232352   0.910470360  0.501281159   0   0
#>   0.501281159   0.962898569  0.465876423   0   0
#>   0.465876423   1.017095507  0.476388188   0   0
#>   0.938774197   0.912478840  0.874976797   0   0
#>   0.874976797   8.696013469  6.086515720   0   1
#>   0.336912912   0.458354021  0.261700847   0   0
#>   0.261700847   0.588699080  0.245668055   0   0
#>   0.245668055   0.596569103  0.255806864   0   0
#>   0.255806864   0.601674592  0.274025281   0   0
#>   0.274025281   0.505992846  0.361108972   0   0
#>   0.361108972   0.486253495  0.336912912   0   0
#>   0.200862777   0.264563882  0.317112134   0   0
#>   0.317112134   0.252907572  0.309806980   0   0
#>   0.309806980   0.247883123  0.151964976   0   0
#>   0.792429809   0.633930626  0.771545833   0   0
#>   0.771545833   0.653346071  0.722969396   0   0
#>   0.722969396   0.781612035  0.643884638   0   0
#>   0.643884638   0.806489135  0.682453674   0   0
#>   0.682453674   0.797473238  0.709742323   0   0
#>   0.709742323   0.661004542  0.812862683   0   0
#>   0.812862683   0.634567717  0.792429809   0   0
#>   0.895884461   0.488351052  0.895233901   0   0
#>   0.895233901   0.482040662  0.852424422   0   0
#>   0.852424422   0.505665831  0.828789202   0   0
#>   0.413810847   1.388908431  0.566901827   0   0
#>   0.566901827   0.925228124  0.569936310   0   0
#>   0.188731581   0.733468869  0.168143863   0   0
#>   0.212175996   0.752991339  0.244836201   0   0
#>   0.244836201   0.728773872  0.188731581   0   0
#>   0.817580007   0.817800381  0.943340716   0   0
#>   0.698906591   0.916442830  0.795783865   0   0
#>   0.795783865   0.901013463  0.815749043   0   0
#>   0.815749043   0.823344057  0.815399912   0   0
#>   0.335055296   0.851780602  0.363331856   0   0
#>   0.363331856   0.832703497  0.397054330   0   0
#>   0.397054330   0.813201862  0.418379240   0   0
#>   0.605674440   0.789481949  0.572143154   0   0
#>   0.572143154   0.856990130  0.535273605   0   0
#>   0.606227013   0.781326371  0.642120391   0   0
#>   0.642120391   0.772216227  0.605674440   0   0
#>   0.390994929   0.902527109  0.464184598   0   0
#>   0.686295010   0.372473417  0.719321781   0   0
#>   0.450159817  -0.103127092  0.377580123   0   0
#>   0.377580123   0.006701703  0.326457890   0   0
#>   0.326457890   0.009551780  0.326737174   0   0
#>   0.326737174   0.157425188  0.366544783   0   0
#>   0.366544783   0.102073586  0.442521306   0   0
#>   0.442521306   0.094220887  0.450159817   0   0
#>   1.991132772   0.533353533  4.543111056   0   0
#>   4.543111056   0.506236071  6.359625540   0   1
#>   0.042335987   0.520232769  0.129256070   0   0
#>   0.600272127   0.648213789  0.657066715   0   0
#>   0.753014085  -0.031108428  0.747804289   0   0
#>   0.747804289   0.103326725  0.796598761   0   0
#>   0.071530730   0.567160292  0.035867597   0   0
#>   0.309339065   0.133983806  0.233231537   0   0
#>   0.683006893   1.192367689  0.797151112   0   0
#>   0.258848106   0.422822718  0.259719724   0   0
#>   0.259719724   0.269487110  0.322855773   0   0
#>   0.232365769   7.344620893  0.181749620   0   1
#>   0.386578355   0.043219035  0.154582187   0   0
#>   0.154582187   0.071159428  0.276018942   0   0
#>   0.516622253  -3.467485839  0.823585086   0   1
#>   0.535591758   0.094248276  0.570272014   0   0
#>   0.570272014   0.072167994  0.583170300   0   0
#>   0.583170300   0.063634009  0.546702318   0   0
#>   0.546702318   0.092674114  0.535591758   0   0
#>   0.012164124  -3.582381398 -9.079237767   0   1
#>   0.938774197  10.054942149  9.437383098   0   1
#>   0.965980047  -1.903294956  4.829035165   0   1
#>   0.753014085  -3.140187465  1.218179843   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.96068100 0.32697426
#>   [2,] 0.29487108 0.61421693
#>   [3,] 0.63017433 0.53115547
#>   [4,] 0.64509129 0.02029297
#>   [5,] 0.26452891 0.76738317
#>   [6,] 0.42925181 0.06910247
#>   [7,] 0.01790102 0.04091916
#>   [8,] 0.09904676 0.04818352
#>   [9,] 0.56290849 0.59957842
#>  [10,] 0.75520792 0.47371409
#>  [11,] 0.48092373 0.46233593
#>  [12,] 0.59136671 0.53050032
#>  [13,] 0.14807511 0.54981046
#>  [14,] 0.74272066 0.97951007
#>  [15,] 0.19694507 0.10807704
#>  [16,] 0.75487813 0.33953571
#>  [17,] 0.25044226 0.01965548
#>  [18,] 0.73540386 0.31173371
#>  [19,] 0.85792369 0.09050501
#>  [20,] 0.96800118 0.50623348
#>  [21,] 0.91567025 0.94135180
#>  [22,] 0.50460303 0.30180607
#>  [23,] 0.29584690 0.19498136
#>  [24,] 0.73198053 0.71551183
#>  [25,] 0.58468734 0.90608534
#>  [26,] 0.98559755 0.52443893
#>  [27,] 0.79179577 0.20728057
#>  [28,] 0.58198510 0.53274422
#>  [29,] 0.53264745 0.72159273
#>  [30,] 0.79196744 0.87661393
#>  [31,] 0.23422817 0.08900089
#>  [32,] 0.71782576 0.09966757
#>  [33,] 0.39924336 0.93342096
#>  [34,] 0.84663786 0.75339346
#>  [35,] 0.76863026 0.36289012
#>  [36,] 0.86377218 0.61395502
#>  [37,] 0.89499171 0.43437392
#>  [38,] 0.34724956 0.66089788
#>  [39,] 0.12366515 0.36600142
#>  [40,] 0.12795485 0.35006649
#>  [41,] 0.31111638 0.99178373
#>  [42,] 0.84607948 0.87761327
#>  [43,] 0.53938992 0.98484200
#>  [44,] 0.51190272 0.04310090
#>  [45,] 0.92737008 0.44606594
#>  [46,] 0.93484418 0.09263485
#>  [47,] 0.94514884 0.02020321
#>  [48,] 0.66752959 0.61408579
#>  [49,] 0.70250012 0.44553169
#>  [50,] 0.11125837 0.88319039
#>  [51,] 0.60731071 0.04416577
#>  [52,] 0.60151501 0.66336625
#>  [53,] 0.20275229 0.19794476
#>  [54,] 0.93028285 0.33591704
#>  [55,] 0.97562598 0.73950744
#>  [56,] 0.54730000 0.00000000
#>  [57,] 0.97506537 0.85264403
#>  [58,] 0.69904635 0.24731571
#>  [59,] 0.35972935 0.23441941
#>  [60,] 0.43559205 0.62550755
#>  [61,] 1.00000000 0.68184962
#>  [62,] 0.05175016 0.21150235
#>  [63,] 0.00000000 0.11657167
#>  [64,] 0.86149855 1.00000000
#>  [65,] 0.80340109 0.26420285
#>  [66,] 0.77403772 0.77116999
#>  [67,] 0.66394106 0.17812350
#>  [68,] 0.16495327 0.73525392
#>  [69,] 0.37633909 0.20939957
#>  [70,] 0.98618786 0.61463967
#>  [71,] 0.07142574 0.55766050
#>  [72,] 0.12911644 0.31037147
#>  [73,] 0.39583621 0.32211050
#>  [74,] 0.56490336 0.36806005
#>  [75,] 0.09402421 0.59634630
#>  [76,] 0.11267169 0.00293590
#>  [77,] 0.86733268 0.62820944
#>  [78,] 0.38743682 0.78117396
#>  [79,] 0.49350537 0.21158316
#>  [80,] 0.69532481 0.65606066
#>  [81,] 0.42833609 0.74980058
#>  [82,] 0.16829109 0.39851301
#>  [83,] 0.98271853 0.43035322
#>  [84,] 0.44944793 0.32226553
#>  [85,] 0.17916117 0.42305358
#>  [86,] 0.58558065 0.07249554
#>  [87,] 0.79445774 0.48704020
#>  [88,] 0.17183037 0.85015500
#>  [89,] 0.06151379 0.55998002
#>  [90,] 0.06042147 0.20950723
#>  [91,] 0.63124334 0.09343351
#>  [92,] 0.98124334 0.07339669
#>  [93,] 0.44817701 0.34783890
#>  [94,] 0.06212154 0.53334174
#>  [95,] 0.44400132 0.98144025
#>  [96,] 0.74428326 0.36340351
#>  [97,] 0.87244321 0.45782708
#>  [98,] 0.54826359 0.97184393
#>  [99,] 0.11583442 0.55564475
#> [100,] 0.71229484 0.53595724
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.960681,0.3269743) [5]: 46 54 65 83 92 
#> 2: (0.2948711,0.6142169) [7]: 5 11 13 38 60 68 85 
#> 3: (0.6301743,0.5311555) [6]: 9 12 48 49 74 100 
#> 4: (0.6450913,0.02029297) [6]: 19 32 47 51 56 91 
#> 5: (0.2645289,0.7673832) [7]: 2 33 38 41 68 78 88 
#> 6: (0.4292518,0.06910247) [7]: 17 23 31 44 56 69 79 
#> 7: (0.01790102,0.04091916) [3]: 8 63 76 
#> 8: (0.09904676,0.04818352) [5]: 7 15 63 76 90 
#> 9: (0.5629085,0.5995784) [8]: 3 11 12 28 29 48 52 60 
#> 10: (0.7552079,0.4737141) [5]: 35 49 87 96 100 
#> 11: (0.4809237,0.4623359) [8]: 2 9 28 60 73 74 85 93 
#> 12: (0.5913667,0.5305003) [4]: 3 9 28 74 
#> 13: (0.1480751,0.5498105) [5]: 2 68 75 85 99 
#> 14: (0.7427207,0.9795101) [5]: 25 30 43 64 98 
#> 15: (0.1969451,0.108077) [6]: 8 17 31 53 76 90 
#> 16: (0.7548781,0.3395357) [4]: 18 35 65 96 
#> 17: (0.2504423,0.01965548) [5]: 6 15 31 56 76 
#> 18: (0.7354039,0.3117337) [5]: 16 58 65 74 96 
#> 19: (0.8579237,0.09050501) [5]: 4 27 32 46 47 
#> 20: (0.9680012,0.5062335) [5]: 26 36 45 83 97 
#> 21: (0.9156702,0.9413518) [3]: 42 57 64 
#> 22: (0.504603,0.3018061) [6]: 58 67 74 79 84 93 
#> 23: (0.2958469,0.1949814) [6]: 6 31 53 59 69 82 
#> 24: (0.7319805,0.7155118) [7]: 25 29 34 52 66 77 80 
#> 25: (0.5846873,0.9060853) [9]: 14 24 29 30 33 66 81 95 98 
#> 26: (0.9855976,0.5244389) [5]: 20 36 61 70 83 
#> 27: (0.7917958,0.2072806) [6]: 19 32 46 58 65 67 
#> 28: (0.5819851,0.5327442) [4]: 9 11 12 74 
#> 29: (0.5326475,0.7215927) [6]: 9 24 25 52 60 81 
#> 30: (0.7919674,0.8766139) [5]: 14 25 42 64 66 
#> 31: (0.2342282,0.08900089) [5]: 6 15 17 23 53 
#> 32: (0.7178258,0.09966757) [5]: 4 19 27 67 91 
#> 33: (0.3992434,0.933421) [6]: 5 25 41 78 81 95 
#> 34: (0.8466379,0.7533935) [6]: 24 42 55 57 66 77 
#> 35: (0.7686303,0.3628901) [8]: 10 16 37 54 65 87 96 97 
#> 36: (0.8637722,0.613955) [8]: 20 26 70 77 80 87 97 100 
#> 37: (0.8949917,0.4343739) [4]: 35 45 54 97 
#> 38: (0.3472496,0.6608979) [5]: 2 5 60 78 81 
#> 39: (0.1236652,0.3660014) [6]: 40 62 72 82 85 94 
#> 40: (0.1279548,0.3500665) [3]: 39 72 82 
#> 41: (0.3111164,0.9917837) [7]: 5 33 43 50 64 88 95 
#> 42: (0.8460795,0.8776133) [6]: 21 30 34 57 64 66 
#> 43: (0.5393899,0.984842) [5]: 14 41 64 95 98 
#> 44: (0.5119027,0.0431009) [4]: 6 56 79 86 
#> 45: (0.9273701,0.4460659) [5]: 20 37 54 83 97 
#> 46: (0.9348442,0.09263485) [6]: 1 19 27 47 65 92 
#> 47: (0.9451488,0.02020321) [5]: 4 19 46 56 92 
#> 48: (0.6675296,0.6140858) [5]: 3 9 52 80 100 
#> 49: (0.7025001,0.4455317) [5]: 3 10 74 96 100 
#> 50: (0.1112584,0.8831904) [5]: 41 68 75 88 89 
#> 51: (0.6073107,0.04416577) [4]: 4 56 86 91 
#> 52: (0.601515,0.6633663) [5]: 9 24 29 48 80 
#> 53: (0.2027523,0.1979448) [6]: 15 23 31 72 82 90 
#> 54: (0.9302828,0.335917) [6]: 1 35 37 45 65 83 
#> 55: (0.975626,0.7395074) [4]: 34 57 61 77 
#> 56: (0.5473,0) [8]: 4 6 17 44 47 51 76 86 
#> 57: (0.9750654,0.852644) [5]: 21 34 42 55 61 
#> 58: (0.6990464,0.2473157) [6]: 18 22 27 65 67 74 
#> 59: (0.3597293,0.2344194) [5]: 23 69 73 82 84 
#> 60: (0.435592,0.6255076) [6]: 2 9 11 29 38 81 
#> 61: (1,0.6818496) [7]: 26 55 57 70 77 83 92 
#> 62: (0.05175016,0.2115023) [5]: 39 63 72 90 94 
#> 63: (0,0.1165717) [6]: 7 8 62 89 90 94 
#> 64: (0.8614986,1) [6]: 14 21 30 41 42 43 
#> 65: (0.8034011,0.2642028) [8]: 1 16 18 27 35 46 54 58 
#> 66: (0.7740377,0.77117) [5]: 24 25 30 34 42 
#> 67: (0.6639411,0.1781235) [6]: 22 27 32 58 79 91 
#> 68: (0.1649533,0.7352539) [6]: 2 5 13 50 75 88 
#> 69: (0.3763391,0.2093996) [5]: 6 23 59 79 84 
#> 70: (0.9861879,0.6146397) [4]: 26 36 61 77 
#> 71: (0.07142574,0.5576605) [4]: 75 89 94 99 
#> 72: (0.1291164,0.3103715) [6]: 39 40 53 62 82 90 
#> 73: (0.3958362,0.3221105) [6]: 11 59 82 84 85 93 
#> 74: (0.5649034,0.36806) [10]: 3 11 12 18 22 28 49 58 93 96 
#> 75: (0.09402421,0.5963463) [6]: 13 50 68 71 89 99 
#> 76: (0.1126717,0.0029359) [5]: 7 8 15 17 56 
#> 77: (0.8673327,0.6282094) [7]: 24 34 36 55 61 70 80 
#> 78: (0.3874368,0.781174) [4]: 5 33 38 81 
#> 79: (0.4935054,0.2115832) [8]: 6 22 44 67 69 84 86 91 
#> 80: (0.6953248,0.6560607) [6]: 24 36 48 52 77 100 
#> 81: (0.4283361,0.7498006) [6]: 25 29 33 38 60 78 
#> 82: (0.1682911,0.398513) [8]: 23 39 40 53 59 72 73 85 
#> 83: (0.9827185,0.4303532) [7]: 1 20 26 45 54 61 92 
#> 84: (0.4494479,0.3222655) [6]: 22 59 69 73 79 93 
#> 85: (0.1791612,0.4230536) [8]: 2 11 13 39 73 82 94 99 
#> 86: (0.5855807,0.07249554) [5]: 44 51 56 79 91 
#> 87: (0.7944577,0.4870402) [5]: 10 35 36 97 100 
#> 88: (0.1718304,0.850155) [4]: 5 41 50 68 
#> 89: (0.06151379,0.55998) [5]: 50 63 71 75 94 
#> 90: (0.06042147,0.2095072) [6]: 8 15 53 62 63 72 
#> 91: (0.6312433,0.09343351) [6]: 4 32 51 67 79 86 
#> 92: (0.9812433,0.07339669) [5]: 1 46 47 61 83 
#> 93: (0.448177,0.3478389) [5]: 11 22 73 74 84 
#> 94: (0.06212154,0.5333417) [7]: 39 62 63 71 85 89 99 
#> 95: (0.4440013,0.9814403) [5]: 25 33 41 43 98 
#> 96: (0.7442833,0.3634035) [6]: 10 16 18 35 49 74 
#> 97: (0.8724432,0.4578271) [6]: 20 35 36 37 45 87 
#> 98: (0.5482636,0.9718439) [4]: 14 25 43 95 
#> 99: (0.1158344,0.5556448) [5]: 13 71 75 85 94 
#> 100: (0.7122948,0.5359572) [7]: 3 10 36 48 49 80 87 
#> number of nodes: 100 
#> number of arcs: 284 
#> number of boundary nodes: 13 
#> boundary nodes:  7 21 41 47 50 56 57 61 63 64 76 89 92 
#> number of triangles: 185 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.050229362 0.356959584 0.108017096 0.145746035 0.216131592 0.186209952 
#>                                                                         
#> 0.130513028 0.304308634 0.169351526 0.095167596 0.134839991 0.104630773 
#>                                                                         
#> 0.157907332 0.090954478 0.090044415 0.105622151 0.074710470 0.164581749 
#>                                                                         
#> 0.398361479 0.112029323 0.140397235 0.086806563 0.054772988 0.074511073 
#>                                                                         
#> 0.198643620 0.115106710 0.126812270 0.152704110 0.105700413 0.081773328 
#>                                                                         
#> 0.122591710 0.071215750 0.297507746 0.055773246 0.169483735 0.117186627 
#>                                                                         
#> 0.047254480 0.120362781 0.134746652 0.170078741 0.110850228 0.082380153 
#>                                                                         
#> 0.090183497 0.136572738 0.123169507 0.009646231 0.095622271 0.092145908 
#>                                                                         
#> 0.024352410 0.126808131 0.165567698 0.239034685 0.058322462 0.118464909 
#>                                                                         
#> 0.155968771 0.046176846 0.146937730 0.322465090 0.015738223 0.075172349 
#>                                                                         
#> 0.209364763 0.141628732 0.138781385 0.113399727 0.142799695 0.052427208 
#>                                                                         
#> 0.082962674 0.104633587 0.434638226 0.179440335 0.180481408 0.105712283 
#>                                                                         
#> 0.073160973 0.183320722 0.130583769 0.094129741 0.151177222 0.156368536 
#>                                                                         
#> 0.056171924 0.057535560 0.123354205 0.131380451 0.025604924 0.119128693 
#>                                                                         
#> 0.173689014 0.077588294 0.030031245 0.252141458 0.203507233 0.191691608 
#>                                                                         
#> 0.094093840 0.174248092 0.126883070 0.074744800 0.232498860 0.104702957 
#>                                                                         
#> 0.065643945 0.185910819 0.108058053 0.106957469 0.252089441 0.068614518 
#>                                                                         
#> 0.122417591 0.095178219 0.124504644 0.128530880 0.090782904 0.133286896 
#>                                                                         
#> 0.051546471 0.155289030 0.113137976 0.104590732 0.032534372 0.083277494 
#>                                                                         
#> 0.121296380 0.119627304 0.110187406 0.160815899 0.322006457 0.125541407 
#>                                                                         
#> 0.039712012 0.063040283 0.026840241 0.096455087 0.068994959 0.095449237 
#>                                                                         
#> 0.081981770 0.166802953 0.064283422 0.048779435 0.050234245 0.122035210 
#>                                                                         
#> 0.172604944 0.143100965 0.134475377 0.053611935 0.240029432 0.421374387 
#>                                                                         
#> 0.026645218 0.181469251 0.058159253 0.254409892 0.105701768 0.031686288 
#>                                                                         
#> 0.169343494 0.235759411 0.160303617 0.223455163 0.240173747 0.141173191 
#>                                                                         
#> 0.070161355 0.156142705 0.177562915 0.112082351 0.082260775 0.090955222 
#>                                                                         
#> 0.095949948 0.038813144 0.175671335 0.300057564 0.224114616 0.107659785 
#>                                                                         
#> 0.074439931 0.044690974 0.099874625 0.124274677 0.123679183 0.213814239 
#>                                                                         
#> 0.229185520 0.183418554 0.196036133 0.185520508 0.136786431 0.086644416 
#>                                                                         
#> 0.156298653 0.149943437 0.081470253 0.077741547 0.114766339 0.165883255 
#>                                                                         
#> 0.074561054 0.125710868 0.129929964 0.159865576 0.069503435 0.075602534 
#>                                                                         
#> 0.059769219 0.110850219 0.111633827 0.041450373 0.164021515 0.119087859 
#>                                                                         
#> 0.126255771 0.071323795 0.032764322 0.203400631 0.194608123 0.174257631 
#>                                                                         
#> 0.114073929 0.120532260 0.103345641 0.041879938 0.090055160 0.033944045 
#>                                                                         
#> 0.089607570 0.027102575 0.026113663 0.179563578 0.073969913 0.076949965 
#>                                                                         
#> 0.134199229 0.025319374 0.149892009 0.107119109 0.072601779 0.077294332 
#>                                                                         
#> 0.094368827 0.106756027 0.058827501 0.090902868 0.201707677 0.201934161 
#>                                                                         
#> 0.089586377 0.072821125 0.075075471 0.240199114 0.093141767 0.199425814 
#>                                                                         
#> 0.140500587 0.069843283 0.161064925 0.120753136 0.069761288 0.240860019 
#>                                                                         
#> 0.159596106 0.187447875 0.221067045 0.158068203 0.090202680 0.131137193 
#>                                                                         
#> 0.058093281 0.101021126 0.054121269 0.129733406 0.162309277 0.124221063 
#>                                                                         
#> 0.163887477 0.145179711 0.140677491 0.170379104 0.144609346 0.014692371 
#>                                                                         
#> 0.173630070 0.034424737 0.120327537 0.178298616 0.170416350 0.055896409 
#>                                                                         
#> 0.016502233 0.055213013 0.079591182 0.228379070 0.550443501 0.079325171 
#>                                                                         
#> 0.050343459 0.287361355 0.157379611 0.035703933 0.134394948 0.062598064 
#>                                                                         
#> 0.125564187 0.094833712 0.608741968 0.108119897 0.008897871 0.447654906 
#>                                                                         
#> 0.044454407 0.044802701 0.010179732 0.026037863 0.102099018 0.079838485 
#>                         
#> 0.227454915 0.327015999 
#> 
#> $alpha
#> [1] 0.2970803
#> 
#> attr(,"class")
#> [1] "scree"