Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2         mx1
#>    87   75 0.008246286 0.082183120 0.238343259 0.058938555  0.06149824
#>    75   58 0.238343259 0.058938555 0.259046863 0.154320947  0.23036824
#>    58   32 0.259046863 0.154320947 0.086500122 0.215845825  0.16330405
#>    32   87 0.086500122 0.215845825 0.008246286 0.082183120  0.10591697
#>    51   21 0.435348309 0.117501202 0.593169978 0.000000000  0.52349861
#>    21   93 0.593169978 0.000000000 0.667548706 0.033829300  0.59470274
#>    93   45 0.667548706 0.033829300 0.624762557 0.191879516  0.59987496
#>    45   56 0.624762557 0.191879516 0.524710241 0.241094450  0.57252986
#>    56   51 0.524710241 0.241094450 0.435348309 0.117501202  0.45870507
#>    38   68 0.821504277 0.841514622 0.809377797 0.883881986  0.79391928
#>    68   83 0.809377797 0.883881986 0.788724712 0.939934900  0.77572203
#>    83   19 0.788724712 0.939934900 0.693685830 0.960076671  0.73425284
#>    19   85 0.693685830 0.960076671 0.721565159 0.835548997  0.70872628
#>    85   38 0.721565159 0.835548997 0.821504277 0.841514622  0.77331704
#>    48   64 0.933620346 0.577164187 0.976511689 0.558171644  0.98427651
#>    64   36 0.976511689 0.558171644 1.000000000 0.470199460  1.49858148
#>    36   55 1.000000000 0.470199460 0.984763917 0.967421396  2.44570953
#>    55  100 0.984763917 0.967421396 0.955992683 0.763519288  1.27108200
#>   100    5 0.955992683 0.763519288 0.958969362 0.709785528  0.96830583
#>     5   48 0.958969362 0.709785528 0.933620346 0.577164187  0.95800189
#>   100   90 0.955992683 0.763519288 0.891045543 0.749489498  0.92815841
#>    90   47 0.891045543 0.749489498 0.886781376 0.721331386  0.92517230
#>    47   48 0.886781376 0.721331386 0.933620346 0.577164187  0.91238293
#>    82   24 0.115568426 0.999133046 0.072249831 0.947969614  0.14007244
#>    24   43 0.072249831 0.947969614 0.055170380 0.890364947  0.12023089
#>    43   72 0.055170380 0.890364947 0.177891092 0.826091617  0.11656297
#>    72   65 0.177891092 0.826091617 0.257752637 0.822143248  0.21908484
#>    65   67 0.257752637 0.822143248 0.419039663 0.972765345  0.29291192
#>    67   82 0.419039663 0.972765345 0.115568426 0.999133046  0.26911005
#>    47   71 0.886781376 0.721331386 0.817793767 0.763146484  0.81162708
#>    71   73 0.817793767 0.763146484 0.699874281 0.719399751  0.78454496
#>    73   17 0.699874281 0.719399751 0.694466129 0.561841504  0.75361468
#>    17   41 0.694466129 0.561841504 0.650185845 0.457253555  0.78995369
#>    41    8 0.650185845 0.457253555 0.774466702 0.320198083  0.79022570
#>     8   46 0.774466702 0.320198083 0.877601550 0.328478684  0.81567274
#>    46   25 0.877601550 0.328478684 0.909775441 0.339249092  0.85185200
#>    25   48 0.909775441 0.339249092 0.933620346 0.577164187  0.86838442
#>    63   39 0.735395279 0.296044607 0.767123194 0.225643834  0.77698881
#>    39   46 0.767123194 0.225643834 0.877601550 0.328478684  0.83053490
#>    41   27 0.650185845 0.457253555 0.632794196 0.400280212  0.72752169
#>    27   63 0.632794196 0.400280212 0.735395279 0.296044607  0.71274474
#>    99   89 0.163576409 0.350564484 0.299865517 0.273442392  0.26030100
#>    89   33 0.299865517 0.273442392 0.355082338 0.218045172  0.34774230
#>    33   23 0.355082338 0.218045172 0.389712022 0.304358745  0.35792069
#>    23   97 0.389712022 0.304358745 0.393658591 0.472563207  0.38505029
#>    97   98 0.393658591 0.472563207 0.337961170 0.469111615  0.37063045
#>    98   99 0.337961170 0.469111615 0.163576409 0.350564484  0.26544214
#>    84   22 0.379416347 0.008614427 0.438883050 0.012224889  0.41192658
#>    22   50 0.438883050 0.012224889 0.426424427 0.027162114  0.42545477
#>    50   28 0.426424427 0.027162114 0.386580900 0.081341456  0.40297948
#>    28   84 0.386580900 0.081341456 0.379416347 0.008614427  0.38657402
#>    81   77 0.204930744 0.672157564 0.160117042 0.639808360  0.22200448
#>    77   59 0.160117042 0.639808360 0.176333118 0.523550693  0.22311297
#>    59   78 0.176333118 0.523550693 0.251794281 0.508357811  0.22805826
#>    78   12 0.251794281 0.508357811 0.416861020 0.527925808  0.32835061
#>    12   42 0.416861020 0.527925808 0.384705038 0.656900795  0.33477606
#>    42   40 0.384705038 0.656900795 0.335315174 0.760383789  0.32791402
#>    40   81 0.335315174 0.760383789 0.204930744 0.672157564  0.26115979
#>    78   98 0.251794281 0.508357811 0.337961170 0.469111615  0.32894654
#>    97   15 0.393658591 0.472563207 0.468461280 0.466378176  0.43265512
#>    15   96 0.468461280 0.466378176 0.465951684 0.633743623  0.50647199
#>    96   34 0.465951684 0.633743623 0.440056859 0.634546325  0.45121074
#>    34   42 0.440056859 0.634546325 0.384705038 0.656900795  0.38973895
#>    55   26 0.984763917 0.967421396 0.849261870 0.992175349  0.91869003
#>    26   37 0.849261870 0.992175349 0.824456003 0.917268696  0.88421962
#>    37   79 0.824456003 0.917268696 0.877682857 0.823669710  0.88733920
#>    79  100 0.877682857 0.823669710 0.955992683 0.763519288  0.95605573
#>    67   53 0.419039663 0.972765345 0.531768453 0.883427513  0.47658493
#>    53   69 0.531768453 0.883427513 0.596515677 0.973625679  0.52347416
#>    69   66 0.596515677 0.973625679 0.632405834 0.997101556  0.54860294
#>    97   91 0.393658591 0.472563207 0.498760052 0.400267430  0.42588351
#>    91   41 0.498760052 0.400267430 0.650185845 0.457253555  0.55964253
#>    41   86 0.650185845 0.457253555 0.637005207 0.548375915  0.56078069
#>    86   44 0.637005207 0.548375915 0.546579955 0.639662902  0.53699087
#>    44   96 0.546579955 0.639662902 0.465951684 0.633743623  0.51257667
#>    45   63 0.624762557 0.191879516 0.735395279 0.296044607  0.63888452
#>    27   91 0.632794196 0.400280212 0.498760052 0.400267430  0.56578048
#>    91   23 0.498760052 0.400267430 0.389712022 0.304358745  0.46788596
#>    23   56 0.389712022 0.304358745 0.524710241 0.241094450  0.46529897
#>    73   86 0.699874281 0.719399751 0.637005207 0.548375915  0.64527911
#>    61   31 0.374971601 0.840565531 0.337590540 0.793310268  0.38485147
#>    31   40 0.337590540 0.793310268 0.335315174 0.760383789  0.38606037
#>    44   54 0.546579955 0.639662902 0.607983789 0.807916943  0.52305747
#>    54   53 0.607983789 0.807916943 0.531768453 0.883427513  0.52034812
#>    53   61 0.531768453 0.883427513 0.374971601 0.840565531  0.44825582
#>    83   26 0.788724712 0.939934900 0.849261870 0.992175349  0.75927330
#>    26   66 0.849261870 0.992175349 0.632405834 0.997101556  0.74349397
#>    69   54 0.596515677 0.973625679 0.607983789 0.807916943  0.63480666
#>    54   85 0.607983789 0.807916943 0.721565159 0.835548997  0.64943515
#>    75   84 0.238343259 0.058938555 0.379416347 0.008614427  0.29165750
#>    28   35 0.386580900 0.081341456 0.357023983 0.164374378  0.34499341
#>    35   57 0.357023983 0.164374378 0.301150340 0.173562587  0.32192667
#>    57   58 0.301150340 0.173562587 0.259046863 0.154320947  0.29862924
#>    70   93 0.890487925 0.115178734 0.667548706 0.033829300  1.48849015
#>    51   22 0.435348309 0.117501202 0.438883050 0.012224889  0.52090451
#>    51   50 0.435348309 0.117501202 0.426424427 0.027162114  0.48889399
#>    33   56 0.355082338 0.218045172 0.524710241 0.241094450  0.43924224
#>    82   76 0.115568426 0.999133046 0.000000000 0.935032679  0.01903486
#>    76   43 0.000000000 0.935032679 0.055170380 0.890364947  0.03880159
#>    70   36 0.890487925 0.115178734 1.000000000 0.470199460  1.25712283
#>    36   48 1.000000000 0.470199460 0.933620346 0.577164187  0.86970061
#>    46   70 0.877601550 0.328478684 0.890487925 0.115178734  0.93025834
#>    83   37 0.788724712 0.939934900 0.824456003 0.917268696  0.82553796
#>    51   52 0.435348309 0.117501202 0.398901562 0.137199149  0.40579712
#>    52   35 0.398901562 0.137199149 0.357023983 0.164374378  0.35624564
#>    87   92 0.008246286 0.082183120 0.042714400 0.277474351 -0.35584738
#>    92   74 0.042714400 0.277474351 0.052411204 0.338538349  0.01795205
#>    74   99 0.052411204 0.338538349 0.163576409 0.350564484  0.09858895
#>    99   59 0.163576409 0.350564484 0.176333118 0.523550693  0.09888698
#>    77   62 0.160117042 0.639808360 0.085632948 0.735035749 -0.04101495
#>    62   94 0.085632948 0.735035749 0.037718631 0.796282058 -0.17653417
#>    94   76 0.037718631 0.796282058 0.000000000 0.935032679 -0.97798270
#>    76   87 0.000000000 0.935032679 0.008246286 0.082183120 -3.32689881
#>    65   81 0.257752637 0.822143248 0.204930744 0.672157564  0.22401362
#>    61   65 0.374971601 0.840565531 0.257752637 0.822143248  0.30580666
#>    58   88 0.259046863 0.154320947 0.249506509 0.221686271  0.16953552
#>    88   95 0.249506509 0.221686271 0.110979974 0.327705594  0.16663638
#>    95   92 0.110979974 0.327705594 0.042714400 0.277474351  0.09967086
#>    52   56 0.398901562 0.137199149 0.524710241 0.241094450  0.44215577
#>    89   57 0.299865517 0.273442392 0.301150340 0.173562587  0.30522079
#>    68   79 0.809377797 0.883881986 0.877682857 0.823669710  0.86545949
#>    85   71 0.721565159 0.835548997 0.817793767 0.763146484  0.77356542
#>    71   79 0.817793767 0.763146484 0.877682857 0.823669710  0.83967982
#>    63   80 0.735395279 0.296044607 0.678832254 0.178134835  0.70296582
#>    80   70 0.678832254 0.178134835 0.890487925 0.115178734  0.77328190
#>    94   72 0.037718631 0.796282058 0.177891092 0.826091617  0.10320983
#>    73   54 0.699874281 0.719399751 0.607983789 0.807916943  0.60557702
#>    93   60 0.667548706 0.033829300 0.663092040 0.179238058  0.61737940
#>    60   63 0.663092040 0.179238058 0.735395279 0.296044607  0.66757076
#>    90   71 0.891045543 0.749489498 0.817793767 0.763146484  0.85157585
#>    67   61 0.419039663 0.972765345 0.374971601 0.840565531  0.44558613
#>    89   88 0.299865517 0.273442392 0.249506509 0.221686271  0.29942966
#>    99   78 0.163576409 0.350564484 0.251794281 0.508357811  0.19776621
#>    93   80 0.667548706 0.033829300 0.678832254 0.178134835  0.66590626
#>    81   72 0.204930744 0.672157564 0.177891092 0.826091617  0.16702002
#>    90   79 0.891045543 0.749489498 0.877682857 0.823669710  0.85921704
#>    95   99 0.110979974 0.327705594 0.163576409 0.350564484  0.09906869
#>    87   84 0.008246286 0.082183120 0.379416347 0.008614427  0.04479155
#>     1   75 0.198306420 0.069311907 0.238343259 0.058938555  0.06149824
#>     1   58 0.198306420 0.069311907 0.259046863 0.154320947  0.23036824
#>     1   32 0.198306420 0.069311907 0.086500122 0.215845825  0.16330405
#>     1   87 0.198306420 0.069311907 0.008246286 0.082183120  0.10591697
#>     2   21 0.535503004 0.170023240 0.593169978 0.000000000  0.52349861
#>     2   93 0.535503004 0.170023240 0.667548706 0.033829300  0.59470274
#>     2   45 0.535503004 0.170023240 0.624762557 0.191879516  0.59987496
#>     2   56 0.535503004 0.170023240 0.524710241 0.241094450  0.57252986
#>     2   51 0.535503004 0.170023240 0.435348309 0.117501202  0.45870507
#>     3   68 0.764046156 0.866246013 0.809377797 0.883881986  0.79391928
#>     3   83 0.764046156 0.866246013 0.788724712 0.939934900  0.77572203
#>     3   19 0.764046156 0.866246013 0.693685830 0.960076671  0.73425284
#>     3   85 0.764046156 0.866246013 0.721565159 0.835548997  0.70872628
#>     3   38 0.764046156 0.866246013 0.821504277 0.841514622  0.77331704
#>     4   64 0.971601610 0.708429884 0.976511689 0.558171644  0.98427651
#>     4   36 0.971601610 0.708429884 1.000000000 0.470199460  1.49858148
#>     4   55 0.971601610 0.708429884 0.984763917 0.967421396  2.44570953
#>     4  100 0.971601610 0.708429884 0.955992683 0.763519288  1.27108200
#>     4    5 0.971601610 0.708429884 0.958969362 0.709785528  0.96830583
#>     4   48 0.971601610 0.708429884 0.933620346 0.577164187  0.95800189
#>     5   90 0.958969362 0.709785528 0.891045543 0.749489498  0.92815841
#>     5   47 0.958969362 0.709785528 0.886781376 0.721331386  0.92517230
#>     6   24 0.182409340 0.879786793 0.072249831 0.947969614  0.14007244
#>     6   43 0.182409340 0.879786793 0.055170380 0.890364947  0.12023089
#>     6   72 0.182409340 0.879786793 0.177891092 0.826091617  0.11656297
#>     6   65 0.182409340 0.879786793 0.257752637 0.822143248  0.21908484
#>     6   67 0.182409340 0.879786793 0.419039663 0.972765345  0.29291192
#>     6   82 0.182409340 0.879786793 0.115568426 0.999133046  0.26911005
#>     7   71 0.838051847 0.591001335 0.817793767 0.763146484  0.81162708
#>     7   73 0.838051847 0.591001335 0.699874281 0.719399751  0.78454496
#>     7   17 0.838051847 0.591001335 0.694466129 0.561841504  0.75361468
#>     7   41 0.838051847 0.591001335 0.650185845 0.457253555  0.78995369
#>     7    8 0.838051847 0.591001335 0.774466702 0.320198083  0.79022570
#>     7   46 0.838051847 0.591001335 0.877601550 0.328478684  0.81567274
#>     7   25 0.838051847 0.591001335 0.909775441 0.339249092  0.85185200
#>     7   48 0.838051847 0.591001335 0.933620346 0.577164187  0.86838442
#>     7   47 0.838051847 0.591001335 0.886781376 0.721331386  0.89453423
#>     8   39 0.774466702 0.320198083 0.767123194 0.225643834  0.77698881
#>     8   27 0.774466702 0.320198083 0.632794196 0.400280212  0.72752169
#>     8   63 0.774466702 0.320198083 0.735395279 0.296044607  0.71274474
#>     9   89 0.345018268 0.314329617 0.299865517 0.273442392  0.26030100
#>     9   33 0.345018268 0.314329617 0.355082338 0.218045172  0.34774230
#>     9   23 0.345018268 0.314329617 0.389712022 0.304358745  0.35792069
#>     9   97 0.345018268 0.314329617 0.393658591 0.472563207  0.38505029
#>     9   98 0.345018268 0.314329617 0.337961170 0.469111615  0.37063045
#>     9   99 0.345018268 0.314329617 0.163576409 0.350564484  0.26544214
#>    10   22 0.413176979 0.019321014 0.438883050 0.012224889  0.41192658
#>    10   50 0.413176979 0.019321014 0.426424427 0.027162114  0.42545477
#>    10   28 0.413176979 0.019321014 0.386580900 0.081341456  0.40297948
#>    10   84 0.413176979 0.019321014 0.379416347 0.008614427  0.38657402
#>    11   77 0.275754459 0.650530836 0.160117042 0.639808360  0.22200448
#>    11   59 0.275754459 0.650530836 0.176333118 0.523550693  0.22311297
#>    11   78 0.275754459 0.650530836 0.251794281 0.508357811  0.22805826
#>    11   12 0.275754459 0.650530836 0.416861020 0.527925808  0.32835061
#>    11   42 0.275754459 0.650530836 0.384705038 0.656900795  0.33477606
#>    11   40 0.275754459 0.650530836 0.335315174 0.760383789  0.32791402
#>    11   81 0.275754459 0.650530836 0.204930744 0.672157564  0.26115979
#>    12   98 0.416861020 0.527925808 0.337961170 0.469111615  0.32894654
#>    12   97 0.416861020 0.527925808 0.393658591 0.472563207  0.36288701
#>    12   15 0.416861020 0.527925808 0.468461280 0.466378176  0.43265512
#>    12   96 0.416861020 0.527925808 0.465951684 0.633743623  0.50647199
#>    12   34 0.416861020 0.527925808 0.440056859 0.634546325  0.45121074
#>    13   26 0.946300361 0.925196836 0.849261870 0.992175349  0.91869003
#>    13   37 0.946300361 0.925196836 0.824456003 0.917268696  0.88421962
#>    13   79 0.946300361 0.925196836 0.877682857 0.823669710  0.88733920
#>    13  100 0.946300361 0.925196836 0.955992683 0.763519288  0.95605573
#>    13   55 0.946300361 0.925196836 0.984763917 0.967421396  1.06965311
#>    14   67 0.461826591 1.000000000 0.419039663 0.972765345  0.28813143
#>    14   53 0.461826591 1.000000000 0.531768453 0.883427513  0.47658493
#>    14   69 0.461826591 1.000000000 0.596515677 0.973625679  0.52347416
#>    14   66 0.461826591 1.000000000 0.632405834 0.997101556  0.54860294
#>    15   91 0.468461280 0.466378176 0.498760052 0.400267430  0.42588351
#>    15   41 0.468461280 0.466378176 0.650185845 0.457253555  0.55964253
#>    15   86 0.468461280 0.466378176 0.637005207 0.548375915  0.56078069
#>    15   44 0.468461280 0.466378176 0.546579955 0.639662902  0.53699087
#>    16   63 0.542172014 0.293229112 0.735395279 0.296044607  0.63888452
#>    16   27 0.542172014 0.293229112 0.632794196 0.400280212  0.63865544
#>    16   91 0.542172014 0.293229112 0.498760052 0.400267430  0.56578048
#>    16   23 0.542172014 0.293229112 0.389712022 0.304358745  0.46788596
#>    16   56 0.542172014 0.293229112 0.524710241 0.241094450  0.46529897
#>    16   45 0.542172014 0.293229112 0.624762557 0.191879516  0.59029596
#>    17   86 0.694466129 0.561841504 0.637005207 0.548375915  0.64527911
#>    18   31 0.431916098 0.798761434 0.337590540 0.793310268  0.38485147
#>    18   40 0.431916098 0.798761434 0.335315174 0.760383789  0.38606037
#>    18   42 0.431916098 0.798761434 0.384705038 0.656900795  0.40354089
#>    18   34 0.431916098 0.798761434 0.440056859 0.634546325  0.44112995
#>    18   96 0.431916098 0.798761434 0.465951684 0.633743623  0.45559205
#>    18   44 0.431916098 0.798761434 0.546579955 0.639662902  0.49965769
#>    18   54 0.431916098 0.798761434 0.607983789 0.807916943  0.52305747
#>    18   53 0.431916098 0.798761434 0.531768453 0.883427513  0.52034812
#>    18   61 0.431916098 0.798761434 0.374971601 0.840565531  0.44825582
#>    19   26 0.693685830 0.960076671 0.849261870 0.992175349  0.75927330
#>    19   66 0.693685830 0.960076671 0.632405834 0.997101556  0.74349397
#>    19   69 0.693685830 0.960076671 0.596515677 0.973625679  0.64184461
#>    19   54 0.693685830 0.960076671 0.607983789 0.807916943  0.63480666
#>    20   84 0.309795017 0.074418947 0.379416347 0.008614427  0.29165750
#>    20   28 0.309795017 0.074418947 0.386580900 0.081341456  0.35086886
#>    20   35 0.309795017 0.074418947 0.357023983 0.164374378  0.34499341
#>    20   57 0.309795017 0.074418947 0.301150340 0.173562587  0.32192667
#>    20   58 0.309795017 0.074418947 0.259046863 0.154320947  0.29862924
#>    20   75 0.309795017 0.074418947 0.238343259 0.058938555  0.26623857
#>    21   22 0.593169978 0.000000000 0.438883050 0.012224889  0.52090451
#>    21   84 0.593169978 0.000000000 0.379416347 0.008614427  0.45566415
#>    24   76 0.072249831 0.947969614 0.000000000 0.935032679  0.01903486
#>    25   36 0.909775441 0.339249092 1.000000000 0.470199460  1.25712283
#>    25   70 0.909775441 0.339249092 0.890487925 0.115178734  0.93025834
#>    28   51 0.386580900 0.081341456 0.435348309 0.117501202  0.43106388
#>    28   52 0.386580900 0.081341456 0.398901562 0.137199149  0.40579712
#>    29   92 0.029242330 0.354267407 0.042714400 0.277474351 -0.35584738
#>    29   74 0.029242330 0.354267407 0.052411204 0.338538349  0.01795205
#>    29   99 0.029242330 0.354267407 0.163576409 0.350564484  0.09858895
#>    29   59 0.029242330 0.354267407 0.176333118 0.523550693  0.09888698
#>    29   77 0.029242330 0.354267407 0.160117042 0.639808360 -0.02974374
#>    29   62 0.029242330 0.354267407 0.085632948 0.735035749 -0.04101495
#>    29   94 0.029242330 0.354267407 0.037718631 0.796282058 -0.17653417
#>    29   76 0.029242330 0.354267407 0.000000000 0.935032679 -0.97798270
#>    29   87 0.029242330 0.354267407 0.008246286 0.082183120 -3.32689881
#>    30   81 0.267072795 0.817019634 0.204930744 0.672157564  0.22401362
#>    30   40 0.267072795 0.817019634 0.335315174 0.760383789  0.25707954
#>    30   31 0.267072795 0.817019634 0.337590540 0.793310268  0.29380163
#>    30   61 0.267072795 0.817019634 0.374971601 0.840565531  0.31679249
#>    30   65 0.267072795 0.817019634 0.257752637 0.822143248  0.30580666
#>    32   88 0.086500122 0.215845825 0.249506509 0.221686271  0.16953552
#>    32   95 0.086500122 0.215845825 0.110979974 0.327705594  0.16663638
#>    32   92 0.086500122 0.215845825 0.042714400 0.277474351  0.09967086
#>    33   57 0.355082338 0.218045172 0.301150340 0.173562587  0.30522079
#>    33   35 0.355082338 0.218045172 0.357023983 0.164374378  0.33260516
#>    33   52 0.355082338 0.218045172 0.398901562 0.137199149  0.40535143
#>    37   68 0.824456003 0.917268696 0.809377797 0.883881986  0.79507046
#>    38   71 0.821504277 0.841514622 0.817793767 0.763146484  0.77356542
#>    38   79 0.821504277 0.841514622 0.877682857 0.823669710  0.83967982
#>    39   80 0.767123194 0.225643834 0.678832254 0.178134835  0.70296582
#>    39   70 0.767123194 0.225643834 0.890487925 0.115178734  0.77328190
#>    43   94 0.055170380 0.890364947 0.037718631 0.796282058 -0.01878679
#>    44   73 0.546579955 0.639662902 0.699874281 0.719399751  0.64189273
#>    45   60 0.624762557 0.191879516 0.663092040 0.179238058  0.61737940
#>    49   54 0.710738831 0.735291812 0.607983789 0.807916943  0.67246079
#>    49   73 0.710738831 0.735291812 0.699874281 0.719399751  0.65323874
#>    49   71 0.710738831 0.735291812 0.817793767 0.763146484  0.78394574
#>    49   85 0.710738831 0.735291812 0.721565159 0.835548997  0.75596572
#>    57   88 0.301150340 0.173562587 0.249506509 0.221686271  0.29942966
#>    60   80 0.663092040 0.179238058 0.678832254 0.178134835  0.66590626
#>    62   81 0.085632948 0.735035749 0.204930744 0.672157564  0.14649373
#>    62   72 0.085632948 0.735035749 0.177891092 0.826091617  0.16702002
#>    74   95 0.052411204 0.338538349 0.110979974 0.327705594  0.07620890
#>    88   99 0.249506509 0.221686271 0.163576409 0.350564484  0.22346101
#>    14   82 0.461826591 1.000000000 0.115568426 0.999133046  0.28813143
#>    21   70 0.593169978 0.000000000 0.890487925 0.115178734  1.48849015
#>          my1         mx2         my2 bp1 bp2
#>  -0.54116041  0.04479155 -0.70653929   0   0
#>   0.11060775  0.26623857  0.10282178   0   0
#>   0.15852629  0.16953552  0.17600247   0   0
#>   0.11473959 -0.01852605  0.18759569   0   0
#>   0.07116059  0.52090451  0.06767633   0   0
#>   0.09531097  1.48849015 -1.86981309   0   0
#>   0.10032566  0.61737940  0.10506432   0   0
#>   0.21200117  0.59029596  0.24811905   0   0
#>   0.19471592  0.45498644  0.19740461   0   0
#>   0.85653830  0.86352867  0.87646205   0   0
#>   0.90331263  0.79507046  0.91044169   0   0
#>   0.91720075  0.75927330  1.03525971   0   0
#>   0.89805928  0.64943515  0.88478515   0   0
#>   0.80867340  0.77356542  0.80451248   0   0
#>   0.63363472  0.92375830  0.49696490   0   0
#>   0.65044098  0.92375830  0.49696490   0   0
#>   0.76334390  4.45499724  0.82491334   0   1
#>   0.82304010  1.06965311  0.85146235   0   0
#>   0.73725207  0.92815841  0.73502803   0   0
#>   0.64123720  0.91238293  0.64995673   0   0
#>   0.73502803  0.91579908  0.79224222   0   0
#>   0.72991953  0.85157585  0.74106472   0   0
#>   0.64995673  0.89453423  0.64415780   0   0
#>   0.93446619  0.01903486  1.03694520   0   0
#>   0.90240920  0.03880159  0.92655251   0   0
#>   0.85828983  0.10320983  0.83279391   0   0
#>   0.84966300  0.21431310  0.75314753   0   0
#>   0.94615907  0.29941339  0.93919726   0   0
#>   1.00673483  0.28813143  1.22565582   0   0
#>   0.67515622  0.85157585  0.74106472   0   0
#>   0.67196919  0.78394574  0.67358439   0   0
#>   0.63868318  0.64527911  0.64240178   0   0
#>   0.45974649  0.67686275  0.50762676   0   0
#>   0.45936442  0.72752169  0.40250491   0   0
#>   0.45338941  0.83053490  0.26828129   0   0
#>   0.45883991  0.93025834  0.22462067   0   0
#>   0.46354995  0.86970061  0.46341804   0   0
#>   0.27243991  0.70296582  0.23907955   0   0
#>   0.26828129  0.87432047  0.22124122   0   0
#>   0.40250491  0.56577220  0.45188046   0   0
#>   0.37636315  0.63865544  0.30343565   0   0
#>   0.36250969  0.22346101  0.29740656   0   0
#>   0.26594615  0.30522079  0.22356311   0   0
#>   0.26701004  0.43924224  0.23438320   0   0
#>   0.38861665  0.41288105  0.38796366   0   0
#>   0.39304925  0.36288701  0.51800292   0   0
#>   0.38825332  0.25566346  0.40263792   0   0
#>  -0.03531745  0.45566415 -0.75570413   0   0
#>   0.01368909  0.48889399  0.06660155   0   0
#>   0.05166083  0.43106388  0.07231412   0   0
#>   0.04462572  0.35086886  0.04814314   0   0
#>   0.60129008  0.14649373  0.70589594   0   0
#>   0.58933550 -0.02974374  0.55406606   0   0
#>   0.58546350  0.19776621  0.43500666   0   0
#>   0.56856140  0.32894654  0.56353442   0   0
#>   0.57595648  0.38973895  0.58965980   0   0
#>   0.69332363  0.40354089  0.72941846   0   0
#>   0.72951682  0.25707954  0.73554678   0   0
#>   0.56353442  0.25566346  0.40263792   0   0
#>   0.48876315  0.42588351  0.40686626   0   0
#>   0.55064967  0.51257667  0.55074121   0   0
#>   0.57628630  0.45559205  0.71762580   0   0
#>   0.58965980  0.44112995  0.71690886   0   0
#>   0.98897893  2.22334500  8.13060263   0   1
#>   0.93903820  0.82553796  0.95847104   0   0
#>   0.89109470  0.86545949  0.87865239   0   0
#>   0.84465236  0.91579908  0.79224222   0   0
#>   0.92958648  0.44558613  0.89047140   0   0
#>   0.95771936  0.61546424  0.89168584   0   0
#>   1.08604771  0.64184461  0.94349891   0   0
#>   0.40686626  0.41288105  0.38796366   0   0
#>   0.46816846  0.56577220  0.45188046   0   0
#>   0.49083576  0.67686275  0.50762676   0   0
#>   0.53973502  0.64189273  0.64364662   0   0
#>   0.55074121  0.49965769  0.72671452   0   0
#>   0.28771423  0.66757076  0.25724685   0   0
#>   0.36512663  0.56577220  0.45188046   0   0
#>   0.32542316  0.41288105  0.38796366   0   0
#>   0.28998504  0.43924224  0.23438320   0   0
#>   0.64240178  0.64189273  0.64364662   0   0
#>   0.79433741  0.31679249  0.84817516   0   0
#>   0.77341893  0.29380163  0.77979442   0   0
#>   0.74357896  0.60557702  0.71346369   0   0
#>   0.79568196  0.61546424  0.89168584   0   0
#>   0.88070523  0.44558613  0.89047140   0   0
#>   1.03525971  0.82553796  0.95847104   0   0
#>   1.11173906  0.84819370  5.72071488   0   1
#>   0.89302446  0.61546424  0.89168584   0   0
#>   0.88478515  0.67246079  0.79013842   0   0
#>  -0.01450261  0.04479155 -0.70653929   0   0
#>   0.11331481  0.35624564  0.11732022   0   0
#>   0.12542545  0.33260516  0.19036150   0   0
#>   0.12339406  0.26820093  0.18997558   0   0
#>  -1.86981309  0.77030293  0.09838864   0   0
#>   0.06767633  0.48889399  0.06660155   0   0
#>   0.06660155  0.43106388  0.07231412   0   0
#>   0.23438320  0.44215577  0.21294150   0   0
#>   1.03694520 -3.65120088  7.65412057   0   1
#>   0.92655251 -0.01878679  0.85542347   0   0
#>   0.19648485  3.82912900 -0.59689343   0   1
#>   0.46341804  0.92375830  0.49696490   0   0
#>   0.22462067  0.87432047  0.22124122   0   0
#>   0.95847104  0.79507046  0.91044169   0   0
#>   0.10639053  0.45498644  0.19740461   0   0
#>   0.11732022  0.40535143  0.19299323   0   0
#>   0.24713154 -0.01852605  0.18759569   0   0
#>   0.31270846  0.07620890  0.30345742   0   0
#>   0.43148648  0.09906869  0.42705199   0   0
#>   0.44229842  0.19776621  0.43500666   0   0
#>   0.55923210  0.14649373  0.70589594   0   0
#>   0.57930208  0.10952608  0.80309322   0   0
#>   0.59467107 -0.01878679  0.85542347   0   0
#>   0.47639992 -4.49932894  0.46506358   0   1
#>   0.74973120  0.21431310  0.75314753   0   0
#>   0.89851760  0.29941339  0.93919726   0   0
#>   0.17600247  0.26820093  0.18997558   0   0
#>   0.25691700  0.17070211  0.26222934   0   0
#>   0.27157201  0.07620890  0.30345742   0   0
#>   0.21294150  0.45498644  0.19740461   0   0
#>   0.22356311  0.29942966  0.22348862   0   0
#>   0.87865239  0.86352867  0.87646205   0   0
#>   0.80451248  0.75596572  0.78112110   0   0
#>   0.80138215  0.85921704  0.78204964   0   0
#>   0.23907955  0.67609843  0.25196823   0   0
#>   0.10840382  0.77030293  0.09838864   0   0
#>   0.83279391  0.10952608  0.80309322   0   0
#>   0.71346369  0.65323874  0.76294175   0   0
#>   0.10506432  0.66590626  0.10655164   0   0
#>   0.25724685  0.67609843  0.25196823   0   0
#>   0.74106472  0.85921704  0.78204964   0   0
#>   0.89047140  0.29941339  0.93919726   0   0
#>   0.22348862  0.22346101  0.29740656   0   0
#>   0.43500666  0.25566346  0.40263792   0   0
#>   0.10655164  0.77030293  0.09838864   0   0
#>   0.74484015  0.21431310  0.75314753   0   0
#>   0.78204964  0.91579908  0.79224222   0   0
#>   0.42705199  0.17070211  0.26222934   0   0
#>  -0.70653929 -0.46902963 -3.29887889   0   1
#>  -0.54116041  0.23036824  0.11060775   0   0
#>   0.11060775  0.16330405  0.15852629   0   0
#>   0.15852629  0.10591697  0.11473959   0   0
#>   0.11473959  0.06149824 -0.54116041   0   0
#>   0.07116059  0.59470274  0.09531097   0   0
#>   0.09531097  0.59987496  0.10032566   0   0
#>   0.10032566  0.57252986  0.21200117   0   0
#>   0.21200117  0.45870507  0.19471592   0   0
#>   0.19471592  0.52349861  0.07116059   0   0
#>   0.85653830  0.77572203  0.90331263   0   0
#>   0.90331263  0.73425284  0.91720075   0   0
#>   0.91720075  0.70872628  0.89805928   0   0
#>   0.89805928  0.77331704  0.80867340   0   0
#>   0.80867340  0.79391928  0.85653830   0   0
#>   0.63363472  1.49858148  0.65044098   0   0
#>   0.65044098  2.44570953  0.76334390   0   0
#>   0.76334390  1.27108200  0.82304010   0   0
#>   0.82304010  0.96830583  0.73725207   0   0
#>   0.73725207  0.95800189  0.64123720   0   0
#>   0.64123720  0.98427651  0.63363472   0   0
#>   0.73502803  0.92517230  0.72991953   0   0
#>   0.72991953  0.91238293  0.64995673   0   0
#>   0.93446619  0.12023089  0.90240920   0   0
#>   0.90240920  0.11656297  0.85828983   0   0
#>   0.85828983  0.21908484  0.84966300   0   0
#>   0.84966300  0.29291192  0.94615907   0   0
#>   0.94615907  0.26911005  1.00673483   0   0
#>   1.00673483  0.14007244  0.93446619   0   0
#>   0.67515622  0.78454496  0.67196919   0   0
#>   0.67196919  0.75361468  0.63868318   0   0
#>   0.63868318  0.78995369  0.45974649   0   0
#>   0.45974649  0.79022570  0.45936442   0   0
#>   0.45936442  0.81567274  0.45338941   0   0
#>   0.45338941  0.85185200  0.45883991   0   0
#>   0.45883991  0.86838442  0.46354995   0   0
#>   0.46354995  0.89453423  0.64415780   0   0
#>   0.64415780  0.81162708  0.67515622   0   0
#>   0.27243991  0.83053490  0.26828129   0   0
#>   0.40250491  0.71274474  0.37636315   0   0
#>   0.37636315  0.77698881  0.27243991   0   0
#>   0.36250969  0.34774230  0.26594615   0   0
#>   0.26594615  0.35792069  0.26701004   0   0
#>   0.26701004  0.38505029  0.38861665   0   0
#>   0.38861665  0.37063045  0.39304925   0   0
#>   0.39304925  0.26544214  0.38825332   0   0
#>   0.38825332  0.26030100  0.36250969   0   0
#>  -0.03531745  0.42545477  0.01368909   0   0
#>   0.01368909  0.40297948  0.05166083   0   0
#>   0.05166083  0.38657402  0.04462572   0   0
#>   0.04462572  0.41192658 -0.03531745   0   0
#>   0.60129008  0.22311297  0.58933550   0   0
#>   0.58933550  0.22805826  0.58546350   0   0
#>   0.58546350  0.32835061  0.56856140   0   0
#>   0.56856140  0.33477606  0.57595648   0   0
#>   0.57595648  0.32791402  0.69332363   0   0
#>   0.69332363  0.26115979  0.72951682   0   0
#>   0.72951682  0.22200448  0.60129008   0   0
#>   0.56353442  0.36288701  0.51800292   0   0
#>   0.51800292  0.43265512  0.48876315   0   0
#>   0.48876315  0.50647199  0.55064967   0   0
#>   0.55064967  0.45121074  0.57628630   0   0
#>   0.57628630  0.38973895  0.58965980   0   0
#>   0.98897893  0.88421962  0.93903820   0   0
#>   0.93903820  0.88733920  0.89109470   0   0
#>   0.89109470  0.95605573  0.84465236   0   0
#>   0.84465236  1.06965311  0.85146235   0   0
#>   0.85146235  0.91869003  0.98897893   0   0
#>   1.22565582  0.47658493  0.92958648   0   0
#>   0.92958648  0.52347416  0.95771936   0   0
#>   0.95771936  0.54860294  1.08604771   0   0
#>   1.08604771  0.64818639  6.94673359   0   1
#>   0.40686626  0.55964253  0.46816846   0   0
#>   0.46816846  0.56078069  0.49083576   0   0
#>   0.49083576  0.53699087  0.53973502   0   0
#>   0.53973502  0.51257667  0.55074121   0   0
#>   0.28771423  0.63865544  0.30343565   0   0
#>   0.30343565  0.56578048  0.36512663   0   0
#>   0.36512663  0.46788596  0.32542316   0   0
#>   0.32542316  0.46529897  0.28998504   0   0
#>   0.28998504  0.59029596  0.24811905   0   0
#>   0.24811905  0.63888452  0.28771423   0   0
#>   0.64240178  0.67686275  0.50762676   0   0
#>   0.79433741  0.38606037  0.77341893   0   0
#>   0.77341893  0.40354089  0.72941846   0   0
#>   0.72941846  0.44112995  0.71690886   0   0
#>   0.71690886  0.45559205  0.71762580   0   0
#>   0.71762580  0.49965769  0.72671452   0   0
#>   0.72671452  0.52305747  0.74357896   0   0
#>   0.74357896  0.52034812  0.79568196   0   0
#>   0.79568196  0.44825582  0.88070523   0   0
#>   0.88070523  0.38485147  0.79433741   0   0
#>   1.03525971  0.74349397  1.11173906   0   0
#>   1.11173906  0.64184461  0.94349891   0   0
#>   0.94349891  0.63480666  0.89302446   0   0
#>   0.89302446  0.64943515  0.88478515   0   0
#>  -0.01450261  0.35086886  0.04814314   0   0
#>   0.04814314  0.34499341  0.11331481   0   0
#>   0.11331481  0.32192667  0.12542545   0   0
#>   0.12542545  0.29862924  0.12339406   0   0
#>   0.12339406  0.26623857  0.10282178   0   0
#>   0.10282178  0.29165750 -0.01450261   0   0
#>   0.06767633  0.45566415 -0.75570413   0   0
#>  -0.75570413  0.26743167 -5.42640132   0   1
#>   1.03694520  0.03880159  0.92655251   0   0
#>   0.19648485  0.86970061  0.46341804   0   0
#>   0.22462067  1.25712283  0.19648485   0   0
#>   0.07231412  0.40579712  0.10639053   0   0
#>   0.10639053  0.35624564  0.11732022   0   0
#>   0.24713154  0.01795205  0.31270846   0   0
#>   0.31270846  0.09858895  0.43148648   0   0
#>   0.43148648  0.09888698  0.44229842   0   0
#>   0.44229842 -0.02974374  0.55406606   0   0
#>   0.55406606 -0.04101495  0.55923210   0   0
#>   0.55923210 -0.17653417  0.57930208   0   0
#>   0.57930208 -0.97798270  0.59467107   0   0
#>   0.59467107 -3.32689881  0.47639992   0   0
#>   0.47639992 -0.35584738  0.24713154   0   0
#>   0.74973120  0.25707954  0.73554678   0   0
#>   0.73554678  0.29380163  0.77979442   0   0
#>   0.77979442  0.31679249  0.84817516   0   0
#>   0.84817516  0.30580666  0.89851760   0   0
#>   0.89851760  0.22401362  0.74973120   0   0
#>   0.17600247  0.16663638  0.25691700   0   0
#>   0.25691700  0.09967086  0.27157201   0   0
#>   0.27157201 -0.01852605  0.18759569   0   0
#>   0.22356311  0.33260516  0.19036150   0   0
#>   0.19036150  0.40535143  0.19299323   0   0
#>   0.19299323  0.44215577  0.21294150   0   0
#>   0.91044169  0.86545949  0.87865239   0   0
#>   0.80451248  0.83967982  0.80138215   0   0
#>   0.80138215  0.86352867  0.87646205   0   0
#>   0.23907955  0.77328190  0.10840382   0   0
#>   0.10840382  0.87432047  0.22124122   0   0
#>   0.85542347  0.10320983  0.83279391   0   0
#>   0.64364662  0.60557702  0.71346369   0   0
#>   0.10506432  0.66757076  0.25724685   0   0
#>   0.79013842  0.65323874  0.76294175   0   0
#>   0.76294175  0.78394574  0.67358439   0   0
#>   0.67358439  0.75596572  0.78112110   0   0
#>   0.78112110  0.67246079  0.79013842   0   0
#>   0.22348862  0.26820093  0.18997558   0   0
#>   0.10655164  0.67609843  0.25196823   0   0
#>   0.70589594  0.16702002  0.74484015   0   0
#>   0.74484015  0.10952608  0.80309322   0   0
#>   0.30345742  0.09906869  0.42705199   0   0
#>   0.29740656  0.17070211  0.26222934   0   0
#>   1.22565582  0.28090051  4.11365626   0   1
#>  -1.86981309  2.62142431 -4.79432568   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.198306420 0.069311907
#>   [2,] 0.535503004 0.170023240
#>   [3,] 0.764046156 0.866246013
#>   [4,] 0.971601610 0.708429884
#>   [5,] 0.958969362 0.709785528
#>   [6,] 0.182409340 0.879786793
#>   [7,] 0.838051847 0.591001335
#>   [8,] 0.774466702 0.320198083
#>   [9,] 0.345018268 0.314329617
#>  [10,] 0.413176979 0.019321014
#>  [11,] 0.275754459 0.650530836
#>  [12,] 0.416861020 0.527925808
#>  [13,] 0.946300361 0.925196836
#>  [14,] 0.461826591 1.000000000
#>  [15,] 0.468461280 0.466378176
#>  [16,] 0.542172014 0.293229112
#>  [17,] 0.694466129 0.561841504
#>  [18,] 0.431916098 0.798761434
#>  [19,] 0.693685830 0.960076671
#>  [20,] 0.309795017 0.074418947
#>  [21,] 0.593169978 0.000000000
#>  [22,] 0.438883050 0.012224889
#>  [23,] 0.389712022 0.304358745
#>  [24,] 0.072249831 0.947969614
#>  [25,] 0.909775441 0.339249092
#>  [26,] 0.849261870 0.992175349
#>  [27,] 0.632794196 0.400280212
#>  [28,] 0.386580900 0.081341456
#>  [29,] 0.029242330 0.354267407
#>  [30,] 0.267072795 0.817019634
#>  [31,] 0.337590540 0.793310268
#>  [32,] 0.086500122 0.215845825
#>  [33,] 0.355082338 0.218045172
#>  [34,] 0.440056859 0.634546325
#>  [35,] 0.357023983 0.164374378
#>  [36,] 1.000000000 0.470199460
#>  [37,] 0.824456003 0.917268696
#>  [38,] 0.821504277 0.841514622
#>  [39,] 0.767123194 0.225643834
#>  [40,] 0.335315174 0.760383789
#>  [41,] 0.650185845 0.457253555
#>  [42,] 0.384705038 0.656900795
#>  [43,] 0.055170380 0.890364947
#>  [44,] 0.546579955 0.639662902
#>  [45,] 0.624762557 0.191879516
#>  [46,] 0.877601550 0.328478684
#>  [47,] 0.886781376 0.721331386
#>  [48,] 0.933620346 0.577164187
#>  [49,] 0.710738831 0.735291812
#>  [50,] 0.426424427 0.027162114
#>  [51,] 0.435348309 0.117501202
#>  [52,] 0.398901562 0.137199149
#>  [53,] 0.531768453 0.883427513
#>  [54,] 0.607983789 0.807916943
#>  [55,] 0.984763917 0.967421396
#>  [56,] 0.524710241 0.241094450
#>  [57,] 0.301150340 0.173562587
#>  [58,] 0.259046863 0.154320947
#>  [59,] 0.176333118 0.523550693
#>  [60,] 0.663092040 0.179238058
#>  [61,] 0.374971601 0.840565531
#>  [62,] 0.085632948 0.735035749
#>  [63,] 0.735395279 0.296044607
#>  [64,] 0.976511689 0.558171644
#>  [65,] 0.257752637 0.822143248
#>  [66,] 0.632405834 0.997101556
#>  [67,] 0.419039663 0.972765345
#>  [68,] 0.809377797 0.883881986
#>  [69,] 0.596515677 0.973625679
#>  [70,] 0.890487925 0.115178734
#>  [71,] 0.817793767 0.763146484
#>  [72,] 0.177891092 0.826091617
#>  [73,] 0.699874281 0.719399751
#>  [74,] 0.052411204 0.338538349
#>  [75,] 0.238343259 0.058938555
#>  [76,] 0.000000000 0.935032679
#>  [77,] 0.160117042 0.639808360
#>  [78,] 0.251794281 0.508357811
#>  [79,] 0.877682857 0.823669710
#>  [80,] 0.678832254 0.178134835
#>  [81,] 0.204930744 0.672157564
#>  [82,] 0.115568426 0.999133046
#>  [83,] 0.788724712 0.939934900
#>  [84,] 0.379416347 0.008614427
#>  [85,] 0.721565159 0.835548997
#>  [86,] 0.637005207 0.548375915
#>  [87,] 0.008246286 0.082183120
#>  [88,] 0.249506509 0.221686271
#>  [89,] 0.299865517 0.273442392
#>  [90,] 0.891045543 0.749489498
#>  [91,] 0.498760052 0.400267430
#>  [92,] 0.042714400 0.277474351
#>  [93,] 0.667548706 0.033829300
#>  [94,] 0.037718631 0.796282058
#>  [95,] 0.110979974 0.327705594
#>  [96,] 0.465951684 0.633743623
#>  [97,] 0.393658591 0.472563207
#>  [98,] 0.337961170 0.469111615
#>  [99,] 0.163576409 0.350564484
#> [100,] 0.955992683 0.763519288
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.1983064,0.06931191) [4]: 32 58 75 87 
#> 2: (0.535503,0.1700232) [5]: 21 45 51 56 93 
#> 3: (0.7640462,0.866246) [5]: 19 38 68 83 85 
#> 4: (0.9716016,0.7084299) [6]: 5 36 48 55 64 100 
#> 5: (0.9589694,0.7097855) [5]: 4 47 48 90 100 
#> 6: (0.1824093,0.8797868) [6]: 24 43 65 67 72 82 
#> 7: (0.8380518,0.5910013) [9]: 8 17 25 41 46 47 48 71 73 
#> 8: (0.7744667,0.3201981) [6]: 7 27 39 41 46 63 
#> 9: (0.3450183,0.3143296) [6]: 23 33 89 97 98 99 
#> 10: (0.413177,0.01932101) [4]: 22 28 50 84 
#> 11: (0.2757545,0.6505308) [7]: 12 40 42 59 77 78 81 
#> 12: (0.416861,0.5279258) [8]: 11 15 34 42 78 96 97 98 
#> 13: (0.9463004,0.9251968) [5]: 26 37 55 79 100 
#> 14: (0.4618266,1) [5]: 53 66 67 69 82 
#> 15: (0.4684613,0.4663782) [7]: 12 41 44 86 91 96 97 
#> 16: (0.542172,0.2932291) [6]: 23 27 45 56 63 91 
#> 17: (0.6944661,0.5618415) [4]: 7 41 73 86 
#> 18: (0.4319161,0.7987614) [9]: 31 34 40 42 44 53 54 61 96 
#> 19: (0.6936858,0.9600767) [7]: 3 26 54 66 69 83 85 
#> 20: (0.309795,0.07441895) [6]: 28 35 57 58 75 84 
#> 21: (0.59317,0) [6]: 2 22 51 70 84 93 
#> 22: (0.4388831,0.01222489) [5]: 10 21 50 51 84 
#> 23: (0.389712,0.3043587) [6]: 9 16 33 56 91 97 
#> 24: (0.07224983,0.9479696) [4]: 6 43 76 82 
#> 25: (0.9097754,0.3392491) [5]: 7 36 46 48 70 
#> 26: (0.8492619,0.9921753) [6]: 13 19 37 55 66 83 
#> 27: (0.6327942,0.4002802) [5]: 8 16 41 63 91 
#> 28: (0.3865809,0.08134146) [7]: 10 20 35 50 51 52 84 
#> 29: (0.02924233,0.3542674) [9]: 59 62 74 76 77 87 92 94 99 
#> 30: (0.2670728,0.8170196) [5]: 31 40 61 65 81 
#> 31: (0.3375905,0.7933103) [4]: 18 30 40 61 
#> 32: (0.08650012,0.2158458) [6]: 1 58 87 88 92 95 
#> 33: (0.3550823,0.2180452) [7]: 9 23 35 52 56 57 89 
#> 34: (0.4400569,0.6345463) [4]: 12 18 42 96 
#> 35: (0.357024,0.1643744) [5]: 20 28 33 52 57 
#> 36: (1,0.4701995) [6]: 4 25 48 55 64 70 
#> 37: (0.824456,0.9172687) [5]: 13 26 68 79 83 
#> 38: (0.8215043,0.8415146) [5]: 3 68 71 79 85 
#> 39: (0.7671232,0.2256438) [5]: 8 46 63 70 80 
#> 40: (0.3353152,0.7603838) [6]: 11 18 30 31 42 81 
#> 41: (0.6501858,0.4572536) [7]: 7 8 15 17 27 86 91 
#> 42: (0.384705,0.6569008) [5]: 11 12 18 34 40 
#> 43: (0.05517038,0.8903649) [5]: 6 24 72 76 94 
#> 44: (0.54658,0.6396629) [6]: 15 18 54 73 86 96 
#> 45: (0.6247626,0.1918795) [6]: 2 16 56 60 63 93 
#> 46: (0.8776016,0.3284787) [5]: 7 8 25 39 70 
#> 47: (0.8867814,0.7213314) [5]: 5 7 48 71 90 
#> 48: (0.9336203,0.5771642) [7]: 4 5 7 25 36 47 64 
#> 49: (0.7107388,0.7352918) [4]: 54 71 73 85 
#> 50: (0.4264244,0.02716211) [4]: 10 22 28 51 
#> 51: (0.4353483,0.1175012) [7]: 2 21 22 28 50 52 56 
#> 52: (0.3989016,0.1371991) [5]: 28 33 35 51 56 
#> 53: (0.5317685,0.8834275) [6]: 14 18 54 61 67 69 
#> 54: (0.6079838,0.8079169) [8]: 18 19 44 49 53 69 73 85 
#> 55: (0.9847639,0.9674214) [5]: 4 13 26 36 100 
#> 56: (0.5247102,0.2410945) [7]: 2 16 23 33 45 51 52 
#> 57: (0.3011503,0.1735626) [6]: 20 33 35 58 88 89 
#> 58: (0.2590469,0.1543209) [6]: 1 20 32 57 75 88 
#> 59: (0.1763331,0.5235507) [5]: 11 29 77 78 99 
#> 60: (0.663092,0.1792381) [4]: 45 63 80 93 
#> 61: (0.3749716,0.8405655) [6]: 18 30 31 53 65 67 
#> 62: (0.08563295,0.7350357) [5]: 29 72 77 81 94 
#> 63: (0.7353953,0.2960446) [7]: 8 16 27 39 45 60 80 
#> 64: (0.9765117,0.5581716) [3]: 4 36 48 
#> 65: (0.2577526,0.8221432) [6]: 6 30 61 67 72 81 
#> 66: (0.6324058,0.9971016) [4]: 14 19 26 69 
#> 67: (0.4190397,0.9727653) [6]: 6 14 53 61 65 82 
#> 68: (0.8093778,0.883882) [5]: 3 37 38 79 83 
#> 69: (0.5965157,0.9736257) [5]: 14 19 53 54 66 
#> 70: (0.8904879,0.1151787) [7]: 21 25 36 39 46 80 93 
#> 71: (0.8177938,0.7631465) [8]: 7 38 47 49 73 79 85 90 
#> 72: (0.1778911,0.8260916) [6]: 6 43 62 65 81 94 
#> 73: (0.6998743,0.7193998) [7]: 7 17 44 49 54 71 86 
#> 74: (0.0524112,0.3385383) [4]: 29 92 95 99 
#> 75: (0.2383433,0.05893855) [5]: 1 20 58 84 87 
#> 76: (0,0.9350327) [6]: 24 29 43 82 87 94 
#> 77: (0.160117,0.6398084) [5]: 11 29 59 62 81 
#> 78: (0.2517943,0.5083578) [5]: 11 12 59 98 99 
#> 79: (0.8776829,0.8236697) [7]: 13 37 38 68 71 90 100 
#> 80: (0.6788323,0.1781348) [5]: 39 60 63 70 93 
#> 81: (0.2049307,0.6721576) [7]: 11 30 40 62 65 72 77 
#> 82: (0.1155684,0.999133) [5]: 6 14 24 67 76 
#> 83: (0.7887247,0.9399349) [5]: 3 19 26 37 68 
#> 84: (0.3794163,0.008614427) [7]: 10 20 21 22 28 75 87 
#> 85: (0.7215652,0.835549) [6]: 3 19 38 49 54 71 
#> 86: (0.6370052,0.5483759) [5]: 15 17 41 44 73 
#> 87: (0.008246286,0.08218312) [7]: 1 29 32 75 76 84 92 
#> 88: (0.2495065,0.2216863) [6]: 32 57 58 89 95 99 
#> 89: (0.2998655,0.2734424) [5]: 9 33 57 88 99 
#> 90: (0.8910455,0.7494895) [5]: 5 47 71 79 100 
#> 91: (0.4987601,0.4002674) [6]: 15 16 23 27 41 97 
#> 92: (0.0427144,0.2774744) [5]: 29 32 74 87 95 
#> 93: (0.6675487,0.0338293) [6]: 2 21 45 60 70 80 
#> 94: (0.03771863,0.7962821) [5]: 29 43 62 72 76 
#> 95: (0.11098,0.3277056) [5]: 32 74 88 92 99 
#> 96: (0.4659517,0.6337436) [5]: 12 15 18 34 44 
#> 97: (0.3936586,0.4725632) [6]: 9 12 15 23 91 98 
#> 98: (0.3379612,0.4691116) [5]: 9 12 78 97 99 
#> 99: (0.1635764,0.3505645) [9]: 9 29 59 74 78 88 89 95 98 
#> 100: (0.9559927,0.7635193) [6]: 4 5 13 55 79 90 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  14 21 26 36 55 66 70 76 82 84 87 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.23126808 0.09760348 0.18318758 0.15488506 0.19675927 0.08171057 0.16373920 
#>                                                                              
#> 0.11150146 0.15251507 0.04406864 0.05973675 0.09714978 0.12761034 0.10011701 
#>                                                                              
#> 0.04690825 0.09105386 0.49745532 0.20592196 0.05381615 0.13502219 0.06644521 
#>                                                                              
#> 0.02847916 0.15158519 0.06703878 0.06008332 0.13853315 0.07995909 0.22068195 
#>                                                                              
#> 0.30461459 0.08067089 0.12577274 0.15765104 0.11357545 0.18501333 0.10346673 
#>                                                                              
#> 0.03392876 0.23910703 0.07722001 0.15093202 0.05956871 0.14626019 0.15659674 
#>                                                                              
#> 0.07821604 0.09300133 0.16825075 0.05580427 0.21086362 0.05957621 0.01945091 
#>                                                                              
#> 0.06725257 0.07307908 0.05526969 0.11738316 0.07697539 0.16622255 0.13292312 
#>                                                                              
#> 0.11466512 0.15742924 0.09468367 0.07505796 0.16738426 0.02590726 0.05969545 
#>                                                                              
#> 0.13774456 0.07890715 0.10767483 0.09874463 0.14383681 0.11103113 0.04288613 
#>                                                                              
#> 0.12756566 0.16179366 0.09207070 0.12849140 0.08084526 0.15195383 0.13403414 
#>                                                                              
#> 0.14522379 0.14908685 0.18221326 0.06025283 0.03300500 0.17910850 0.10728757 
#>                                                                              
#> 0.16254969 0.07996132 0.21691198 0.16610510 0.11689422 0.14978029 0.08813670 
#>                                                                              
#> 0.05662409 0.04629194 0.23731756 0.10533564 0.09077878 0.17118672 0.13215490 
#>                                                                              
#> 0.07098575 0.37152740 0.12588769 0.21368886 0.04231409 0.04142915 0.04992219 
#>                                                                              
#> 0.19830965 0.06182912 0.11181382 0.17345594 0.12089721 0.07776177 0.14378606 
#>                                                                              
#> 0.85288943 0.15901528 0.11865777 0.06803753 0.17444110 0.08475474 0.16316267 
#>                                                                              
#> 0.09988807 0.09105547 0.12042454 0.08514555 0.13077496 0.22082027 0.14330711 
#>                                                                              
#> 0.12758980 0.14547704 0.13737368 0.07451400 0.13935130 0.07221306 0.18077922 
#>                                                                              
#> 0.14474600 0.15629087 0.07537417 0.05734905 0.37839076 0.04135886 0.10447937 
#>                                                                              
#> 0.18431722 0.19049547 0.17953658 0.18969675 0.09189649 0.07188603 0.11309079 
#>                                                                              
#> 0.04864139 0.07771154 0.11728072 0.05241128 0.06255460 0.15033844 0.23991708 
#>                                                                              
#> 0.25932576 0.05725802 0.01270478 0.13665014 0.07867687 0.07310549 0.12955313 
#>                                                                              
#> 0.12767792 0.05388494 0.09486512 0.25424186 0.13678902 0.17333304 0.18862448 
#>                                                                              
#> 0.14651674 0.23061245 0.27816806 0.26548507 0.26176987 0.09656503 0.13914197 
#>                                                                              
#> 0.09483898 0.16273981 0.04593437 0.06091417 0.09680899 0.04579247 0.16554078 
#>                                                                              
#> 0.15494279 0.18502463 0.02666753 0.01539408 0.06748249 0.03541767 0.11613347 
#>                                                                              
#> 0.16127170 0.14417787 0.18693061 0.10913663 0.12496059 0.07405210 0.09840882 
#>                                                                              
#> 0.06002808 0.08031624 0.11665035 0.10911453 0.11790924 0.12210202 0.12254028 
#>                                                                              
#> 0.16196781 0.05711706 0.05071930 0.13594487 0.13724706 0.17060387 0.07272308 
#>                                                                              
#> 0.18195350 0.18743181 0.19007926 0.19324378 0.14025804 0.11550671 0.15286569 
#>                                                                              
#> 0.05498124 0.13073996 0.05901762 0.09448294 0.10394509 0.14951028 0.16441677 
#>                                                                              
#> 0.16849124 0.19611258 0.17630557 0.13091538 0.07064176 0.15885285 0.07159665 
#>                                                                              
#> 0.09811022 0.17463511 0.09579856 0.07709730 0.10159997 0.09951981 0.09465572 
#>                                                                              
#> 0.07310948 0.15477049 0.21392714 0.07339893 0.15902349 0.22489894 0.06071069 
#>                                                                              
#> 0.05720035 0.07796583 0.02800357 0.13438511 0.22425996 0.31410480 0.38492133 
#>                                                                              
#> 0.44209592 0.58150100 0.27289319 0.15762821 0.08868281 0.07439682 0.11043804 
#>                                                                              
#> 0.01063564 0.16311098 0.11450708 0.07559937 0.06990966 0.05370590 0.09195762 
#>                                                                              
#> 0.03663366 0.07845593 0.05894467 0.10026163 0.16559407 0.09568779 0.17279212 
#>                                                                              
#> 0.04036032 0.12582928 0.01925087 0.11061936 0.10084003 0.07059018 0.01577883 
#>                                                                   
#> 0.13485411 0.12962537 0.05956215 0.15489860 0.34625925 0.31884809 
#> 
#> $alpha
#> [1] 0.2931715
#> 
#> attr(,"class")
#> [1] "scree"