Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1         my1
#>    18   53 0.05684929 0.77416666 0.16902251 0.72701619  0.117201701  0.76073998
#>    18    1 0.05684929 0.77416666 0.07411634 0.80508343 -0.001162272  0.82684641
#>     1   53 0.07411634 0.80508343 0.16902251 0.72701619  0.149919193  0.80051456
#>    32   70 0.22044031 0.66191650 0.27402851 0.72075465  0.221008853  0.71522110
#>    57   32 0.19883291 0.56660089 0.22044031 0.66191650  0.119151621  0.63477103
#>    76    1 0.01279798 0.74973864 0.07411634 0.80508343 -0.045598530  0.87607885
#>     1   12 0.07411634 0.80508343 0.16929000 0.87394288  0.112140362  0.85273033
#>    62   12 0.26919972 0.81855108 0.16929000 0.87394288  0.193845618  0.80043459
#>    70   62 0.27402851 0.72075465 0.26919972 0.81855108  0.224114783  0.76730754
#>    12   83 0.16929000 0.87394288 0.25662291 0.85578112  0.213023239  0.86518313
#>    24   45 0.19052548 0.91875723 0.08954979 0.90934941  0.139921706  0.91529761
#>    24   12 0.19052548 0.91875723 0.16929000 0.87394288  0.191218067  0.89099061
#>    40   83 0.21168723 0.90976501 0.25662291 0.85578112  0.266149187  0.90940467
#>    83   62 0.25662291 0.85578112 0.26919972 0.81855108  0.272503194  0.84040637
#>    21   62 0.28544155 0.85832746 0.26919972 0.81855108  0.286661388  0.83462517
#>    77   80 0.37342912 0.66654812 0.38134973 0.73716598  0.331038708  0.70705581
#>    33   14 0.29649424 0.85622613 0.31625328 0.88805487  0.295121783  0.87912564
#>     8   32 0.28531520 0.65905536 0.22044031 0.66191650  0.250431172  0.60501084
#>    76   57 0.01279798 0.74973864 0.19883291 0.56660089  0.074637805  0.62649891
#>    40   14 0.21168723 0.90976501 0.31625328 0.88805487  0.278056860  0.96675750
#>    83   21 0.25662291 0.85578112 0.28544155 0.85832746  0.266456945  0.90883604
#>    50   14 0.38576843 0.81834060 0.31625328 0.88805487  0.345591848  0.84779420
#>    66   24 0.18604045 0.99544852 0.19052548 0.91875723  0.220101714  0.95896369
#>    50   62 0.38576843 0.81834060 0.26919972 0.81855108  0.327416114  0.78080940
#>    57   16 0.19883291 0.56660089 0.19023559 0.49810795  0.270237297  0.52285208
#>     8   57 0.28531520 0.65905536 0.19883291 0.56660089  0.263342319  0.59293370
#>    59   16 0.05842017 0.46175393 0.19023559 0.49810795  0.106833716  0.54336270
#>     8   77 0.28531520 0.65905536 0.37342912 0.66654812  0.326341471  0.69844231
#>    33   62 0.29649424 0.85622613 0.26919972 0.81855108  0.327459916  0.80506778
#>    66   40 0.18604045 0.99544852 0.21168723 0.90976501  0.277525732  0.97615184
#>    76   59 0.01279798 0.74973864 0.05842017 0.46175393  0.057644889  0.60923717
#>    36    6 0.42920564 0.57535633 0.46378510 0.66760823  0.419021246  0.63178061
#>    77   64 0.37342912 0.66654812 0.33174668 0.57967761  0.356170922  0.62139365
#>    36   19 0.42920564 0.57535633 0.38528195 0.58280195  0.399554419  0.53371747
#>    60   50 0.35042738 0.91760896 0.38576843 0.81834060  0.387853047  0.87500791
#>    85   50 0.47760959 0.85247777 0.38576843 0.81834060  0.452290991  0.77998250
#>    85   71 0.47760959 0.85247777 0.44388853 0.86631570  0.440822162  0.81083774
#>    81   28 0.52595083 0.66300066 0.52827022 0.76885810  0.511001346  0.71628234
#>    77    6 0.37342912 0.66654812 0.46378510 0.66760823  0.418503027  0.67594944
#>    16   87 0.19023559 0.49810795 0.39742628 0.48264437  0.293272204  0.48288997
#>    59   27 0.05842017 0.46175393 0.04450660 0.40032401  0.136823123  0.41170542
#>     5   27 0.00000000 0.41587955 0.04450660 0.40032401  0.031824739  0.43548702
#>    66   60 0.18604045 0.99544852 0.35042738 0.91760896  0.292478423  1.00772996
#>    15   60 0.39905112 0.99667837 0.35042738 0.91760896  0.405088112  0.93848063
#>    19   64 0.38528195 0.58280195 0.33174668 0.57967761  0.361556625  0.52911015
#>    85   74 0.47760959 0.85247777 0.46337455 0.94359077  0.477023275  0.89905467
#>    28   50 0.52827022 0.76885810 0.38576843 0.81834060  0.450917901  0.77602822
#>    60   71 0.35042738 0.91760896 0.44388853 0.86631570  0.410286276  0.91588337
#>    27   99 0.04450660 0.40032401 0.02077116 0.38071851  0.075992819  0.33803473
#>    36   87 0.42920564 0.57535633 0.39742628 0.48264437  0.431770594  0.52267456
#>    75   87 0.38662591 0.43575161 0.39742628 0.48264437  0.293202724  0.48195905
#>    85   28 0.47760959 0.85247777 0.52827022 0.76885810  0.545315625  0.83634109
#>    50   80 0.38576843 0.81834060 0.38134973 0.73716598  0.447987833  0.77424614
#>    81    6 0.52595083 0.66300066 0.46378510 0.66760823  0.490867892  0.61133518
#>    66   15 0.18604045 0.99544852 0.39905112 0.99667837  0.273790313  4.24452143
#>    15   74 0.39905112 0.99667837 0.46337455 0.94359077  0.434138049  0.97367890
#>    42  100 0.49657595 0.99722558 0.46680712 1.00000000  0.479112661  0.97094214
#>    92   75 0.14311444 0.31890296 0.38662591 0.43575161  0.261927801  0.38345917
#>    97   42 0.54527680 0.90594353 0.49657595 0.99722558  0.515263934  0.94856352
#>    63   97 0.56851848 0.90280052 0.54527680 0.90594353  0.547658421  0.83605045
#>    81   35 0.52595083 0.66300066 0.47654800 0.55054806  0.558250764  0.58173245
#>     7   35 0.49549658 0.50560304 0.47654800 0.55054806  0.442780954  0.50984523
#>    81   65 0.52595083 0.66300066 0.59500998 0.77225285  0.564460306  0.71511103
#>    65   63 0.59500998 0.77225285 0.56851848 0.90280052  0.558473978  0.83280049
#>    81   43 0.52595083 0.66300066 0.64090061 0.61031237  0.599131135  0.67092095
#>    55   78 0.49131299 0.45393267 0.55828061 0.40544079  0.557314081  0.47459332
#>     7   87 0.49549658 0.50560304 0.39742628 0.48264437  0.448980138  0.48336478
#>    17   55 0.40027363 0.35619608 0.49131299 0.45393267  0.445893959  0.40497062
#>    92   17 0.14311444 0.31890296 0.40027363 0.35619608  0.266382547  0.37417552
#>     4   88 0.15137273 0.21497601 0.20792116 0.24390424  0.159947161  0.26794894
#>    42   63 0.49657595 0.99722558 0.56851848 0.90280052  0.577469220  0.98423916
#>    22   43 0.67798478 0.56513546 0.64090061 0.61031237  0.588625262  0.52959233
#>    99   90 0.02077116 0.38071851 0.05174926 0.15990143  0.042194646  0.27114250
#>    89   37 0.67130213 0.68539284 0.70633101 0.68627565  0.686971448  0.75904652
#>    78    3 0.55828061 0.40544079 0.59955503 0.40135882  0.587851830  0.49373509
#>    17   72 0.40027363 0.35619608 0.36791239 0.23509557  0.450450830  0.27791327
#>    52   88 0.31105561 0.21655297 0.20792116 0.24390424  0.244743805  0.17463066
#>    52   72 0.31105561 0.21655297 0.36791239 0.23509557  0.310280466  0.31537047
#>    65   89 0.59500998 0.77225285 0.67130213 0.68539284  0.672872788  0.76370743
#>    37   43 0.70633101 0.68627565 0.64090061 0.61031237  0.702706069  0.62323734
#>     3   22 0.59955503 0.40135882 0.67798478 0.56513546  0.597259328  0.50312582
#>    78   93 0.55828061 0.40544079 0.51649272 0.35877566  0.573604590  0.34967564
#>    69    3 0.54281136 0.28865955 0.59955503 0.40135882  0.577671791  0.34174220
#>    90   39 0.05174926 0.15990143 0.17474449 0.04031614  0.136575132  0.12410224
#>    90    4 0.05174926 0.15990143 0.15137273 0.21497601  0.061382578  0.26011677
#>    31   90 0.02407963 0.16085208 0.05174926 0.15990143  0.041717764  0.27107560
#>    34   65 0.63563245 0.92653815 0.59500998 0.77225285  0.637014862  0.84368368
#>     5   31 0.00000000 0.41587955 0.02407963 0.16085208 -0.211424133  0.26726640
#>    93   17 0.51649272 0.35877566 0.40027363 0.35619608  0.459715362  0.29746632
#>    52   96 0.31105561 0.21655297 0.23635621 0.09662820  0.290605050  0.14606436
#>    34   95 0.63563245 0.92653815 0.71829044 0.82752652  0.732926259  0.92375351
#>    95   37 0.71829044 0.82752652 0.70633101 0.68627565  0.755234178  0.75326685
#>    22    9 0.67798478 0.56513546 0.73305523 0.50976423  0.729199170  0.56100038
#>    37   22 0.70633101 0.68627565 0.67798478 0.56513546  0.711738979  0.62112368
#>    69   54 0.54281136 0.28865955 0.52964550 0.21362291  0.542478984  0.25004451
#>    37   47 0.70633101 0.68627565 0.77023210 0.59192468  0.793848151  0.67673375
#>    68   44 0.54066621 0.21147062 0.65969981 0.16888057  0.620824990  0.24786721
#>    69   72 0.54281136 0.28865955 0.36791239 0.23509557  0.454244498  0.26552606
#>    58    3 0.71336714 0.24891512 0.59955503 0.40135882  0.638010363  0.31136195
#>    54   84 0.52964550 0.21362291 0.28560107 0.07283984  0.442356288  0.08302246
#>    54   72 0.52964550 0.21362291 0.36791239 0.23509557  0.432323293  0.10041444
#>    39   84 0.17474449 0.04031614 0.28560107 0.07283984  0.236673682  0.03441975
#>    10   95 0.77310852 0.83510123 0.71829044 0.82752652  0.756200833  0.75531573
#>    10    2 0.77310852 0.83510123 0.83556524 0.81349545  0.814886852  0.85479558
#>    10   98 0.77310852 0.83510123 0.83603891 0.77197338  0.793249982  0.79224900
#>    39   96 0.17474449 0.04031614 0.23635621 0.09662820  0.152729665  0.12626393
#>    42   34 0.49657595 0.99722558 0.63563245 0.92653815  0.675524922  1.17713462
#>    61   34 0.82633629 0.97306127 0.63563245 0.92653815  0.736154417  0.92860707
#>    61   10 0.82633629 0.97306127 0.77310852 0.83510123  0.772828034  0.91445764
#>    98   11 0.83603891 0.77197338 0.87625318 0.70771291  0.911266066  0.77433731
#>    73   13 0.02087720 0.00000000 0.02712207 0.00701722 -0.064373773  0.08215520
#>     2   98 0.83556524 0.81349545 0.83603891 0.77197338  0.921802168  0.79371548
#>    98   37 0.83603891 0.77197338 0.70633101 0.68627565  0.791027468  0.69909187
#>    47    9 0.77023210 0.59192468 0.73305523 0.50976423  0.878507869  0.49343954
#>    39   13 0.17474449 0.04031614 0.02712207 0.00701722  0.125944889 -0.08721604
#>     5   73 0.00000000 0.41587955 0.02087720 0.00000000 -1.804927434  0.11680819
#>     9    3 0.73305523 0.50976423 0.59955503 0.40135882  0.706012684  0.40666206
#>    30   58 0.68826657 0.19112443 0.71336714 0.24891512  0.624453908  0.25318695
#>     9   56 0.73305523 0.50976423 0.77996648 0.32990144  0.797697371  0.43057497
#>    56   58 0.77996648 0.32990144 0.71336714 0.24891512  0.722283647  0.30945985
#>    68   54 0.54066621 0.21147062 0.52964550 0.21362291  0.517599660  0.12265104
#>    26   23 0.74178495 0.24523025 0.73551829 0.24971027  0.725185033  0.22863316
#>    82   44 0.59943038 0.08112133 0.65969981 0.16888057  0.587259102  0.15405496
#>    44   30 0.65969981 0.16888057 0.68826657 0.19112443  0.621003693  0.24804160
#>    56   23 0.77996648 0.32990144 0.73551829 0.24971027  0.765745839  0.28536972
#>    30   26 0.68826657 0.19112443 0.74178495 0.24523025  0.722828472  0.21045934
#>    39   46 0.17474449 0.04031614 0.76611492 0.03311526  0.461087735 -0.73049233
#>    41   82 0.67978587 0.06650706 0.59943038 0.08112133  0.557402995 -0.37818468
#>    49   94 0.97317465 0.90619041 0.84747529 0.88752615  0.914300504  0.87008402
#>    67   25 0.95018015 0.70991366 0.98329800 0.86705378  0.925020078  0.79727615
#>    20   67 0.97825406 0.63254401 0.95018015 0.70991366  0.914868583  0.65332251
#>    20   47 0.97825406 0.63254401 0.77023210 0.59192468  0.896187271  0.49985254
#>    56   26 0.77996648 0.32990144 0.74178495 0.24523025  0.839832944  0.25196097
#>    38   79 0.95100127 0.24139118 0.85855326 0.34844030  0.846415192  0.24451404
#>    49   61 0.97317465 0.90619041 0.82633629 0.97306127  0.902937284  0.94661263
#>    20   25 0.97825406 0.63254401 0.98329800 0.86705378  1.169560131  0.74573844
#>    25   94 0.98329800 0.86705378 0.84747529 0.88752615  0.909697027  0.83954253
#>    41   46 0.67978587 0.06650706 0.76611492 0.03311526  0.733212324  0.07634170
#>    29   82 0.69561031 0.11554877 0.59943038 0.08112133  0.645230956  0.10473091
#>    73   46 0.02087720 0.00000000 0.76611492 0.03311526  0.453689636 -1.33806003
#>    49   25 0.97317465 0.90619041 0.98329800 0.86705378  1.035583144  0.90145581
#>    29   46 0.69561031 0.11554877 0.76611492 0.03311526  0.810840468  0.14273634
#>    29   86 0.69561031 0.11554877 0.92213201 0.10234640  0.811117085  0.14748245
#>    46   82 0.76611492 0.03311526 0.59943038 0.08112133  0.463161727 -0.70540504
#>    48   20 0.99357682 0.23361283 0.97825406 0.63254401  1.048927816  0.43549870
#>    91   20 1.00000000 0.95294441 0.97825406 0.63254401  4.372078573  0.56313941
#>    38   48 0.95100127 0.24139118 0.99357682 0.23361283  0.997390424  0.37489680
#>    86   48 0.92213201 0.10234640 0.99357682 0.23361283  0.959430878  0.16712159
#>    86   38 0.92213201 0.10234640 0.95100127 0.24139118  0.845429393  0.19079122
#>    76   18 0.01279798 0.74973864 0.05684929 0.77416666  0.081972639  0.67692836
#>    12   53 0.16929000 0.87394288 0.16902251 0.72701619  0.149919193  0.80051456
#>    53   70 0.16902251 0.72701619 0.27402851 0.72075465  0.221008853  0.71522110
#>    53   32 0.16902251 0.72701619 0.22044031 0.66191650  0.119151621  0.63477103
#>    45    1 0.08954979 0.90934941 0.07411634 0.80508343 -0.045598530  0.87607885
#>    45   12 0.08954979 0.90934941 0.16929000 0.87394288  0.112140362  0.85273033
#>    53   62 0.16902251 0.72701619 0.26919972 0.81855108  0.224114783  0.76730754
#>    40   12 0.21168723 0.90976501 0.16929000 0.87394288  0.191218067  0.89099061
#>    14   83 0.31625328 0.88805487 0.25662291 0.85578112  0.266149187  0.90940467
#>    70   80 0.27402851 0.72075465 0.38134973 0.73716598  0.331038708  0.70705581
#>    21   14 0.28544155 0.85832746 0.31625328 0.88805487  0.295121783  0.87912564
#>    53   57 0.16902251 0.72701619 0.19883291 0.56660089  0.074637805  0.62649891
#>    60   14 0.35042738 0.91760896 0.31625328 0.88805487  0.278056860  0.96675750
#>    40   24 0.21168723 0.90976501 0.19052548 0.91875723  0.220101714  0.95896369
#>    80   62 0.38134973 0.73716598 0.26919972 0.81855108  0.327416114  0.78080940
#>    64   16 0.33174668 0.57967761 0.19023559 0.49810795  0.270237297  0.52285208
#>    64   57 0.33174668 0.57967761 0.19883291 0.56660089  0.263342319  0.59293370
#>    70   77 0.27402851 0.72075465 0.37342912 0.66654812  0.326341471  0.69844231
#>    64    8 0.33174668 0.57967761 0.28531520 0.65905536  0.331903124  0.63303789
#>    60   40 0.35042738 0.91760896 0.21168723 0.90976501  0.277525732  0.97615184
#>    57   59 0.19883291 0.56660089 0.05842017 0.46175393  0.057644889  0.60923717
#>    51   77 0.45852371 0.69871753 0.37342912 0.66654812  0.410095641  0.69818867
#>    87   19 0.39742628 0.48264437 0.38528195 0.58280195  0.399554419  0.53371747
#>    71   50 0.44388853 0.86631570 0.38576843 0.81834060  0.387853047  0.87500791
#>    51   28 0.45852371 0.69871753 0.52827022 0.76885810  0.511001346  0.71628234
#>    51    6 0.45852371 0.69871753 0.46378510 0.66760823  0.418503027  0.67594944
#>    77   36 0.37342912 0.66654812 0.42920564 0.57535633  0.415848402  0.62983997
#>    64   87 0.33174668 0.57967761 0.39742628 0.48264437  0.293272204  0.48288997
#>    92   27 0.14311444 0.31890296 0.04450660 0.40032401  0.136823123  0.41170542
#>     5   76 0.00000000 0.41587955 0.01279798 0.74973864 -0.086660985  0.58637641
#>    74   60 0.46337455 0.94359077 0.35042738 0.91760896  0.405088112  0.93848063
#>    71   74 0.44388853 0.86631570 0.46337455 0.94359077  0.477023275  0.89905467
#>    51   50 0.45852371 0.69871753 0.38576843 0.81834060  0.450917901  0.77602822
#>    92   99 0.14311444 0.31890296 0.02077116 0.38071851  0.075992819  0.33803473
#>    99    5 0.02077116 0.38071851 0.00000000 0.41587955  0.021023572  0.40458335
#>    35   87 0.47654800 0.55054806 0.39742628 0.48264437  0.431770594  0.52267456
#>    97   28 0.54527680 0.90594353 0.52827022 0.76885810  0.545315625  0.83634109
#>    51   80 0.45852371 0.69871753 0.38134973 0.73716598  0.447987833  0.77424614
#>    35   36 0.47654800 0.55054806 0.42920564 0.57535633  0.477464632  0.60987385
#>    35    6 0.47654800 0.55054806 0.46378510 0.66760823  0.490867892  0.61133518
#>    51   81 0.45852371 0.69871753 0.52595083 0.66300066  0.496636360  0.68916380
#>    55   75 0.49131299 0.45393267 0.38662591 0.43575161  0.438328346  0.44853363
#>    92   59 0.14311444 0.31890296 0.05842017 0.46175393  0.142255808  0.41492638
#>   100   15 0.46680712 1.00000000 0.39905112 0.99667837  0.273790313  4.24452143
#>   100   74 0.46680712 1.00000000 0.46337455 0.94359077  0.434138049  0.97367890
#>    16   75 0.19023559 0.49810795 0.38662591 0.43575161  0.261927801  0.38345917
#>    97   85 0.54527680 0.90594353 0.47760959 0.85247777  0.493704637  0.90166088
#>    74   42 0.46337455 0.94359077 0.49657595 0.99722558  0.515263934  0.94856352
#>    43   35 0.64090061 0.61031237 0.47654800 0.55054806  0.558250764  0.58173245
#>    28   65 0.52827022 0.76885810 0.59500998 0.77225285  0.564460306  0.71511103
#>    28   63 0.52827022 0.76885810 0.56851848 0.90280052  0.558473978  0.83280049
#>    89   43 0.67130213 0.68539284 0.64090061 0.61031237  0.599131135  0.67092095
#>     7   78 0.49549658 0.50560304 0.55828061 0.40544079  0.557314081  0.47459332
#>    55   87 0.49131299 0.45393267 0.39742628 0.48264437  0.448980138  0.48336478
#>    75   17 0.38662591 0.43575161 0.40027363 0.35619608  0.266382547  0.37417552
#>    88   92 0.20792116 0.24390424 0.14311444 0.31890296  0.268219636  0.36150765
#>    34   63 0.63563245 0.92653815 0.56851848 0.90280052  0.577469220  0.98423916
#>    89   81 0.67130213 0.68539284 0.52595083 0.66300066  0.595329433  0.69559842
#>    43    7 0.64090061 0.61031237 0.49549658 0.50560304  0.565567481  0.56161138
#>    92   90 0.14311444 0.31890296 0.05174926 0.15990143  0.042194646  0.27114250
#>     7    3 0.49549658 0.50560304 0.59955503 0.40135882  0.587851830  0.49373509
#>    93   55 0.51649272 0.35877566 0.49131299 0.45393267  0.508849239  0.40766304
#>    96   88 0.23635621 0.09662820 0.20792116 0.24390424  0.244743805  0.17463066
#>    17   52 0.40027363 0.35619608 0.31105561 0.21655297  0.286155640  0.33078383
#>    95   89 0.71829044 0.82752652 0.67130213 0.68539284  0.672872788  0.76370743
#>     7   22 0.49549658 0.50560304 0.67798478 0.56513546  0.597259328  0.50312582
#>     3   93 0.59955503 0.40135882 0.51649272 0.35877566  0.573604590  0.34967564
#>     4   39 0.15137273 0.21497601 0.17474449 0.04031614  0.136575132  0.12410224
#>    92    4 0.14311444 0.31890296 0.15137273 0.21497601  0.061382578  0.26011677
#>    95   65 0.71829044 0.82752652 0.59500998 0.77225285  0.637014862  0.84368368
#>    99   31 0.02077116 0.38071851 0.02407963 0.16085208 -0.211424133  0.26726640
#>    69   17 0.54281136 0.28865955 0.40027363 0.35619608  0.459715362  0.29746632
#>    84   96 0.28560107 0.07283984 0.23635621 0.09662820  0.290605050  0.14606436
#>    47   22 0.77023210 0.59192468 0.67798478 0.56513546  0.711738979  0.62112368
#>    11   47 0.87625318 0.70771291 0.77023210 0.59192468  0.793848151  0.67673375
#>    84   52 0.28560107 0.07283984 0.31105561 0.21655297  0.370087153  0.13198645
#>    96    4 0.23635621 0.09662820 0.15137273 0.21497601  0.211016776  0.16811888
#>    13   31 0.02712207 0.00701722 0.02407963 0.16085208  0.035294684  0.08412637
#>    69   44 0.54281136 0.28865955 0.65969981 0.16888057  0.620824990  0.24786721
#>    82   84 0.59943038 0.08112133 0.28560107 0.07283984  0.442356288  0.08302246
#>    84   72 0.28560107 0.07283984 0.36791239 0.23509557  0.432323293  0.10041444
#>    98   95 0.83603891 0.77197338 0.71829044 0.82752652  0.756200833  0.75531573
#>    94    2 0.84747529 0.88752615 0.83556524 0.81349545  0.814886852  0.85479558
#>    61   42 0.82633629 0.97306127 0.49657595 0.99722558  1.327344429 10.07224788
#>    10   34 0.77310852 0.83510123 0.63563245 0.92653815  0.736154417  0.92860707
#>    94   10 0.84747529 0.88752615 0.77310852 0.83510123  0.772828034  0.91445764
#>    67   11 0.95018015 0.70991366 0.87625318 0.70771291  0.911266066  0.77433731
#>    13   90 0.02712207 0.00701722 0.05174926 0.15990143  0.089256214  0.07543404
#>    67   98 0.95018015 0.70991366 0.83603891 0.77197338  0.921802168  0.79371548
#>    11   37 0.87625318 0.70771291 0.70633101 0.68627565  0.791027468  0.69909187
#>    79    9 0.85855326 0.34844030 0.73305523 0.50976423  0.878507869  0.49343954
#>    31   73 0.02407963 0.16085208 0.02087720 0.00000000 -1.804927434  0.11680819
#>    56    3 0.77996648 0.32990144 0.59955503 0.40135882  0.706012684  0.40666206
#>    69   58 0.54281136 0.28865955 0.71336714 0.24891512  0.624453908  0.25318695
#>    79   56 0.85855326 0.34844030 0.77996648 0.32990144  0.797697371  0.43057497
#>    23   58 0.73551829 0.24971027 0.71336714 0.24891512  0.722283647  0.30945985
#>    82   54 0.59943038 0.08112133 0.52964550 0.21362291  0.517599660  0.12265104
#>    69   30 0.54281136 0.28865955 0.68826657 0.19112443  0.621003693  0.24804160
#>    58   26 0.71336714 0.24891512 0.74178495 0.24523025  0.722828472  0.21045934
#>    29   44 0.69561031 0.11554877 0.65969981 0.16888057  0.694549372  0.15359031
#>    84   46 0.28560107 0.07283984 0.76611492 0.03311526  0.461087735 -0.73049233
#>     2   25 0.83556524 0.81349545 0.98329800 0.86705378  0.925020078  0.79727615
#>    79   47 0.85855326 0.34844030 0.77023210 0.59192468  0.896187271  0.49985254
#>    79   26 0.85855326 0.34844030 0.74178495 0.24523025  0.839832944  0.25196097
#>    29   30 0.69561031 0.11554877 0.68826657 0.19112443  0.772649384  0.16117934
#>    94   61 0.84747529 0.88752615 0.82633629 0.97306127  0.902937284  0.94661263
#>    11   20 0.87625318 0.70771291 0.97825406 0.63254401  0.876410165  0.60113614
#>    91   49 1.00000000 0.95294441 0.97317465 0.90619041  0.914111954  0.97115052
#>    91   25 1.00000000 0.95294441 0.98329800 0.86705378  1.035583144  0.90145581
#>    86   46 0.92213201 0.10234640 0.76611492 0.03311526  0.810840468  0.14273634
#>    26   86 0.74178495 0.24523025 0.92213201 0.10234640  0.811117085  0.14748245
#>    79   20 0.85855326 0.34844030 0.97825406 0.63254401  1.048927816  0.43549870
#>    79   48 0.85855326 0.34844030 0.99357682 0.23361283  0.997390424  0.37489680
#>    26   38 0.74178495 0.24523025 0.95100127 0.24139118  0.845429393  0.19079122
#>    76   53 0.01279798 0.74973864 0.16902251 0.72701619  0.081972639  0.67692836
#>    45   76 0.08954979 0.90934941 0.01279798 0.74973864 -0.045598530  0.87607885
#>    33   21 0.29649424 0.85622613 0.28544155 0.85832746  0.286661388  0.83462517
#>     8   70 0.28531520 0.65905536 0.27402851 0.72075465  0.253967831  0.68520296
#>    33   50 0.29649424 0.85622613 0.38576843 0.81834060  0.345591848  0.84779420
#>    66   45 0.18604045 0.99544852 0.08954979 0.90934941  0.136309930  0.95406341
#>    19   77 0.38528195 0.58280195 0.37342912 0.66654812  0.356170922  0.62139365
#>    59    5 0.05842017 0.46175393 0.00000000 0.41587955  0.031824739  0.43548702
#>    92   16 0.14311444 0.31890296 0.19023559 0.49810795  0.142255808  0.41492638
#>   100   66 0.46680712 1.00000000 0.18604045 0.99544852  0.273790313  4.24452143
#>    97   74 0.54527680 0.90594353 0.46337455 0.94359077  0.493704637  0.90166088
#>     7   55 0.49549658 0.50560304 0.49131299 0.45393267  0.557314081  0.47459332
#>    88   17 0.20792116 0.24390424 0.40027363 0.35619608  0.268219636  0.36150765
#>    93   69 0.51649272 0.35877566 0.54281136 0.28865955  0.577671791  0.34174220
#>    68   69 0.54066621 0.21147062 0.54281136 0.28865955  0.542478984  0.25004451
#>    61  100 0.82633629 0.97306127 0.46680712 1.00000000  1.327344429 10.07224788
#>    67    2 0.95018015 0.70991366 0.83556524 0.81349545  0.921802168  0.79371548
#>    73   39 0.02087720 0.00000000 0.17474449 0.04031614  0.125944889 -0.08721604
#>    82   68 0.59943038 0.08112133 0.54066621 0.21147062  0.517599660  0.12265104
#>    29   26 0.69561031 0.11554877 0.74178495 0.24523025  0.772649384  0.16117934
#>    29   41 0.69561031 0.11554877 0.67978587 0.06650706  0.733212324  0.07634170
#>    91   61 1.00000000 0.95294441 0.82633629 0.97306127  0.914111954  0.97115052
#>    48   91 0.99357682 0.23361283 1.00000000 0.95294441  4.372078573  0.56313941
#>           mx2         my2 bp1 bp2
#>   0.081972639  0.67692836   0   0
#>   0.117201701  0.76073998   0   0
#>   0.117201701  0.76073998   0   0
#>   0.253967831  0.68520296   0   0
#>   0.250431172  0.60501084   0   0
#>  -0.001162272  0.82684641   0   0
#>   0.149919193  0.80051456   0   0
#>   0.202987032  0.81692288   0   0
#>   0.321092916  0.77209593   0   0
#>   0.202987032  0.81692288   0   0
#>   0.136309930  0.95406341   0   0
#>   0.139921706  0.91529761   0   0
#>   0.213023239  0.86518313   0   0
#>   0.202987032  0.81692288   0   0
#>   0.272503194  0.84040637   0   0
#>   0.410095641  0.69818867   0   0
#>   0.345591848  0.84779420   0   0
#>   0.253967831  0.68520296   0   0
#>   0.057644889  0.60923717   0   0
#>   0.266149187  0.90940467   0   0
#>   0.272503194  0.84040637   0   0
#>   0.364571004  0.86671915   0   0
#>   0.136309930  0.95406341   0   0
#>   0.327459916  0.80506778   0   0
#>   0.106833716  0.54336270   0   0
#>   0.250431172  0.60501084   0   0
#>   0.142255808  0.41492638   0   0
#>   0.331903124  0.63303789   0   0
#>   0.286661388  0.83462517   0   0
#>   0.220101714  0.95896369   0   0
#>  -0.086660985  0.58637641   0   0
#>   0.477464632  0.60987385   0   0
#>   0.331903124  0.63303789   0   0
#>   0.415848402  0.62983997   0   0
#>   0.364571004  0.86671915   0   0
#>   0.440822162  0.81083774   0   0
#>   0.477023275  0.89905467   0   0
#>   0.564460306  0.71511103   0   0
#>   0.419021246  0.63178061   0   0
#>   0.293202724  0.48195905   0   0
#>   0.031824739  0.43548702   0   0
#>   0.021023572  0.40458335   0   0
#>   0.277525732  0.97615184   0   0
#>   0.292478423  1.00772996   0   0
#>   0.356170922  0.62139365   0   0
#>   0.493704637  0.90166088   0   0
#>   0.452290991  0.77998250   0   0
#>   0.387853047  0.87500791   0   0
#>   0.021023572  0.40458335   0   0
#>   0.399554419  0.53371747   0   0
#>   0.438328346  0.44853363   0   0
#>   0.452290991  0.77998250   0   0
#>   0.327416114  0.78080940   0   0
#>   0.496636360  0.68916380   0   0
#>   0.292478423  1.00772996   0   0
#>   0.405088112  0.93848063   0   0
#>   1.327344429 10.07224788   0   0
#>   0.266382547  0.37417552   0   0
#>   0.566575804  0.97593946   0   0
#>   0.566575804  0.97593946   0   0
#>   0.490867892  0.61133518   0   0
#>   0.565567481  0.56161138   0   0
#>   0.595329433  0.69559842   0   0
#>   0.626172811  0.84653834   0   0
#>   0.558250764  0.58173245   0   0
#>   0.508849239  0.40766304   0   0
#>   0.442780954  0.50984523   0   0
#>   0.457570622  0.39409408   0   0
#>   0.268219636  0.36150765   0   0
#>   0.211016776  0.16811888   0   0
#>   0.566575804  0.97593946   0   0
#>   0.702706069  0.62323734   0   0
#>   0.041717764  0.27107560   0   0
#>   0.690120967  0.63407742   0   0
#>   0.573604590  0.34967564   0   0
#>   0.310280466  0.31537047   0   0
#>   0.286155640  0.33078383   0   0
#>   0.370087153  0.13198645   0   0
#>   0.595329433  0.69559842   0   0
#>   0.690120967  0.63407742   0   0
#>   0.647039475  0.47928699   0   0
#>   0.508849239  0.40766304   0   0
#>   0.638010363  0.31136195   0   0
#>   0.089256214  0.07543404   0   0
#>   0.136575132  0.12410224   0   0
#>   0.035294684  0.08412637   0   0
#>   0.626172811  0.84653834   0   0
#>  -1.804927434  0.11680819   0   0
#>   0.457570622  0.39409408   0   0
#>   0.244743805  0.17463066   0   0
#>   0.637014862  0.84368368   0   0
#>   0.686971448  0.75904652   0   0
#>   0.647039475  0.47928699   0   0
#>   0.702706069  0.62323734   0   0
#>   0.454244498  0.26552606   0   0
#>   0.711738979  0.62112368   0   0
#>   0.587259102  0.15405496   0   0
#>   0.450450830  0.27791327   0   0
#>   0.680969554  0.34343461   0   0
#>   0.432323293  0.10041444   0   0
#>   0.454244498  0.26552606   0   0
#>   0.461087735 -0.73049233   0   0
#>   0.732926259  0.92375351   0   0
#>   0.793249982  0.79224900   0   0
#>   0.756200833  0.75531573   0   0
#>   0.236673682  0.03441975   0   0
#>   0.577469220  0.98423916   0   0
#>   0.675524922  1.17713462   0   0
#>   0.736154417  0.92860707   0   0
#>   0.791027468  0.69909187   0   0
#>   0.125944889 -0.08721604   0   0
#>   0.793249982  0.79224900   0   0
#>   0.755234178  0.75326685   0   0
#>   0.729199170  0.56100038   0   0
#>   0.089256214  0.07543404   0   0
#>  -4.203425452 -0.00359668   0   1
#>   0.647039475  0.47928699   0   0
#>   0.722828472  0.21045934   0   0
#>   0.706012684  0.40666206   0   0
#>   0.680969554  0.34343461   0   0
#>   0.542478984  0.25004451   0   0
#>   0.765745839  0.28536972   0   0
#>   0.640716652  0.11734252   0   0
#>   0.694549372  0.15359031   0   0
#>   0.722283647  0.30945985   0   0
#>   0.772649384  0.16117934   0   0
#>   0.453689636 -1.33806003   0   0
#>   0.645230956  0.10473091   0   0
#>   0.902937284  0.94661263   0   0
#>   1.169560131  0.74573844   0   0
#>   1.169560131  0.74573844   0   0
#>   0.876410165  0.60113614   0   0
#>   0.765745839  0.28536972   0   0
#>   0.997390424  0.37489680   0   0
#>   0.914111954  0.97115052   0   0
#>   1.919291850  0.72961286   0   0
#>   0.914300504  0.87008402   0   0
#>   0.557402995 -0.37818468   0   0
#>   0.640716652  0.11734252   0   0
#>   0.513198530 -2.67726940   0   1
#>   0.914300504  0.87008402   0   0
#>   0.733212324  0.07634170   0   0
#>   0.810840468  0.14273634   0   0
#>   0.557402995 -0.37818468   0   0
#>   4.372078573  0.56313941   0   0
#>   1.919291850  0.72961286   0   0
#>   0.959430878  0.16712159   0   0
#>   6.836531307 -3.03162727   0   1
#>   0.959430878  0.16712159   0   0
#>  -0.001162272  0.82684641   0   0
#>   0.193845618  0.80043459   0   0
#>   0.224114783  0.76730754   0   0
#>   0.221008853  0.71522110   0   0
#>   0.112140362  0.85273033   0   0
#>   0.139921706  0.91529761   0   0
#>   0.193845618  0.80043459   0   0
#>   0.213023239  0.86518313   0   0
#>   0.266456945  0.90883604   0   0
#>   0.321092916  0.77209593   0   0
#>   0.266456945  0.90883604   0   0
#>   0.119151621  0.63477103   0   0
#>   0.364571004  0.86671915   0   0
#>   0.191218067  0.89099061   0   0
#>   0.321092916  0.77209593   0   0
#>   0.293272204  0.48288997   0   0
#>   0.270237297  0.52285208   0   0
#>   0.331038708  0.70705581   0   0
#>   0.263342319  0.59293370   0   0
#>   0.278056860  0.96675750   0   0
#>   0.106833716  0.54336270   0   0
#>   0.418503027  0.67594944   0   0
#>   0.361556625  0.52911015   0   0
#>   0.440822162  0.81083774   0   0
#>   0.450917901  0.77602822   0   0
#>   0.496636360  0.68916380   0   0
#>   0.419021246  0.63178061   0   0
#>   0.361556625  0.52911015   0   0
#>   0.075992819  0.33803473   0   0
#>  -3.077541662  0.70102730   0   1
#>   0.410286276  0.91588337   0   0
#>   0.410286276  0.91588337   0   0
#>   0.447987833  0.77424614   0   0
#>   0.042194646  0.27114250   0   0
#>  -0.211424133  0.26726640   0   0
#>   0.442780954  0.50984523   0   0
#>   0.547658421  0.83605045   0   0
#>   0.410095641  0.69818867   0   0
#>   0.431770594  0.52267456   0   0
#>   0.477464632  0.60987385   0   0
#>   0.511001346  0.71628234   0   0
#>   0.445893959  0.40497062   0   0
#>   0.136823123  0.41170542   0   0
#>   0.434138049  0.97367890   0   0
#>   0.479112661  0.97094214   0   0
#>   0.293202724  0.48195905   0   0
#>   0.545315625  0.83634109   0   0
#>   0.479112661  0.97094214   0   0
#>   0.565567481  0.56161138   0   0
#>   0.558473978  0.83280049   0   0
#>   0.547658421  0.83605045   0   0
#>   0.690120967  0.63407742   0   0
#>   0.587851830  0.49373509   0   0
#>   0.438328346  0.44853363   0   0
#>   0.445893959  0.40497062   0   0
#>   0.159947161  0.26794894   0   0
#>   0.626172811  0.84653834   0   0
#>   0.599131135  0.67092095   0   0
#>   0.588625262  0.52959233   0   0
#>   0.061382578  0.26011677   0   0
#>   0.597259328  0.50312582   0   0
#>   0.457570622  0.39409408   0   0
#>   0.211016776  0.16811888   0   0
#>   0.310280466  0.31537047   0   0
#>   0.686971448  0.75904652   0   0
#>   0.588625262  0.52959233   0   0
#>   0.577671791  0.34174220   0   0
#>   0.152729665  0.12626393   0   0
#>   0.159947161  0.26794894   0   0
#>   0.672872788  0.76370743   0   0
#>   0.041717764  0.27107560   0   0
#>   0.450450830  0.27791327   0   0
#>   0.236673682  0.03441975   0   0
#>   0.729199170  0.56100038   0   0
#>   0.876410165  0.60113614   0   0
#>   0.290605050  0.14606436   0   0
#>   0.152729665  0.12626393   0   0
#>  -0.064373773  0.08215520   0   0
#>   0.621003693  0.24804160   0   0
#>   0.463161727 -0.70540504   0   0
#>   0.370087153  0.13198645   0   0
#>   0.755234178  0.75326685   0   0
#>   0.909697027  0.83954253   0   0
#>   0.675524922  1.17713462   0   0
#>   0.732926259  0.92375351   0   0
#>   0.814886852  0.85479558   0   0
#>   0.914868583  0.65332251   0   0
#>   0.035294684  0.08412637   0   0
#>   0.911266066  0.77433731   0   0
#>   0.793848151  0.67673375   0   0
#>   0.797697371  0.43057497   0   0
#>  -0.064373773  0.08215520   0   0
#>   0.680969554  0.34343461   0   0
#>   0.638010363  0.31136195   0   0
#>   0.839832944  0.25196097   0   0
#>   0.725185033  0.22863316   0   0
#>   0.442356288  0.08302246   0   0
#>   0.624453908  0.25318695   0   0
#>   0.725185033  0.22863316   0   0
#>   0.640716652  0.11734252   0   0
#>   0.463161727 -0.70540504   0   0
#>   0.909697027  0.83954253   0   0
#>   0.878507869  0.49343954   0   0
#>   0.846415192  0.24451404   0   0
#>   0.694549372  0.15359031   0   0
#>   0.772828034  0.91445764   0   0
#>   0.914868583  0.65332251   0   0
#>   1.035583144  0.90145581   0   0
#>   1.919291850  0.72961286   0   0
#>   3.187117992 -5.21236686   0   1
#>   0.845429393  0.19079122   0   0
#>   0.896187271  0.49985254   0   0
#>   1.048927816  0.43549870   0   0
#>   0.846415192  0.24451404   0   0
#>   0.074637805  0.62649891   0   0
#>  -5.134183027  3.32301947   0   1
#>   0.295121783  0.87912564   0   0
#>   0.326341471  0.69844231   0   0
#>   0.327459916  0.80506778   0   0
#>  -5.012079897  6.72382657   0   1
#>   0.415848402  0.62983997   0   0
#>  -0.086660985  0.58637641   0   0
#>   0.261927801  0.38345917   0   0
#>   0.216067543  7.80526196   0   1
#>   0.515263934  0.94856352   0   0
#>   0.448980138  0.48336478   0   0
#>   0.286155640  0.33078383   0   0
#>   0.459715362  0.29746632   0   0
#>   0.620824990  0.24786721   0   0
#>   1.534586257 12.83813521   0   1
#>   0.925020078  0.79727615   0   0
#>   0.453689636 -1.33806003   0   0
#>   0.587259102  0.15405496   0   0
#>   0.811117085  0.14748245   0   0
#>   0.645230956  0.10473091   0   0
#>   1.572304935  6.65316206   0   1
#>   5.762147216  0.55072698   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.07411634 0.80508343
#>   [2,] 0.83556524 0.81349545
#>   [3,] 0.59955503 0.40135882
#>   [4,] 0.15137273 0.21497601
#>   [5,] 0.00000000 0.41587955
#>   [6,] 0.46378510 0.66760823
#>   [7,] 0.49549658 0.50560304
#>   [8,] 0.28531520 0.65905536
#>   [9,] 0.73305523 0.50976423
#>  [10,] 0.77310852 0.83510123
#>  [11,] 0.87625318 0.70771291
#>  [12,] 0.16929000 0.87394288
#>  [13,] 0.02712207 0.00701722
#>  [14,] 0.31625328 0.88805487
#>  [15,] 0.39905112 0.99667837
#>  [16,] 0.19023559 0.49810795
#>  [17,] 0.40027363 0.35619608
#>  [18,] 0.05684929 0.77416666
#>  [19,] 0.38528195 0.58280195
#>  [20,] 0.97825406 0.63254401
#>  [21,] 0.28544155 0.85832746
#>  [22,] 0.67798478 0.56513546
#>  [23,] 0.73551829 0.24971027
#>  [24,] 0.19052548 0.91875723
#>  [25,] 0.98329800 0.86705378
#>  [26,] 0.74178495 0.24523025
#>  [27,] 0.04450660 0.40032401
#>  [28,] 0.52827022 0.76885810
#>  [29,] 0.69561031 0.11554877
#>  [30,] 0.68826657 0.19112443
#>  [31,] 0.02407963 0.16085208
#>  [32,] 0.22044031 0.66191650
#>  [33,] 0.29649424 0.85622613
#>  [34,] 0.63563245 0.92653815
#>  [35,] 0.47654800 0.55054806
#>  [36,] 0.42920564 0.57535633
#>  [37,] 0.70633101 0.68627565
#>  [38,] 0.95100127 0.24139118
#>  [39,] 0.17474449 0.04031614
#>  [40,] 0.21168723 0.90976501
#>  [41,] 0.67978587 0.06650706
#>  [42,] 0.49657595 0.99722558
#>  [43,] 0.64090061 0.61031237
#>  [44,] 0.65969981 0.16888057
#>  [45,] 0.08954979 0.90934941
#>  [46,] 0.76611492 0.03311526
#>  [47,] 0.77023210 0.59192468
#>  [48,] 0.99357682 0.23361283
#>  [49,] 0.97317465 0.90619041
#>  [50,] 0.38576843 0.81834060
#>  [51,] 0.45852371 0.69871753
#>  [52,] 0.31105561 0.21655297
#>  [53,] 0.16902251 0.72701619
#>  [54,] 0.52964550 0.21362291
#>  [55,] 0.49131299 0.45393267
#>  [56,] 0.77996648 0.32990144
#>  [57,] 0.19883291 0.56660089
#>  [58,] 0.71336714 0.24891512
#>  [59,] 0.05842017 0.46175393
#>  [60,] 0.35042738 0.91760896
#>  [61,] 0.82633629 0.97306127
#>  [62,] 0.26919972 0.81855108
#>  [63,] 0.56851848 0.90280052
#>  [64,] 0.33174668 0.57967761
#>  [65,] 0.59500998 0.77225285
#>  [66,] 0.18604045 0.99544852
#>  [67,] 0.95018015 0.70991366
#>  [68,] 0.54066621 0.21147062
#>  [69,] 0.54281136 0.28865955
#>  [70,] 0.27402851 0.72075465
#>  [71,] 0.44388853 0.86631570
#>  [72,] 0.36791239 0.23509557
#>  [73,] 0.02087720 0.00000000
#>  [74,] 0.46337455 0.94359077
#>  [75,] 0.38662591 0.43575161
#>  [76,] 0.01279798 0.74973864
#>  [77,] 0.37342912 0.66654812
#>  [78,] 0.55828061 0.40544079
#>  [79,] 0.85855326 0.34844030
#>  [80,] 0.38134973 0.73716598
#>  [81,] 0.52595083 0.66300066
#>  [82,] 0.59943038 0.08112133
#>  [83,] 0.25662291 0.85578112
#>  [84,] 0.28560107 0.07283984
#>  [85,] 0.47760959 0.85247777
#>  [86,] 0.92213201 0.10234640
#>  [87,] 0.39742628 0.48264437
#>  [88,] 0.20792116 0.24390424
#>  [89,] 0.67130213 0.68539284
#>  [90,] 0.05174926 0.15990143
#>  [91,] 1.00000000 0.95294441
#>  [92,] 0.14311444 0.31890296
#>  [93,] 0.51649272 0.35877566
#>  [94,] 0.84747529 0.88752615
#>  [95,] 0.71829044 0.82752652
#>  [96,] 0.23635621 0.09662820
#>  [97,] 0.54527680 0.90594353
#>  [98,] 0.83603891 0.77197338
#>  [99,] 0.02077116 0.38071851
#> [100,] 0.46680712 1.00000000
#> 
#> $tri.obj
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,18,53), [2,4,3]
#> 2: (76,53,18), [1,3,23]
#> 3: (76,18,1), [1,7,2]
#> 4: (12,1,53), [1,9,8]
#> 5: (53,32,70), [19,10,6]
#> 6: (53,57,32), [22,5,23]
#> 7: (45,76,1), [3,8,0]
#> 8: (45,1,12), [4,13,7]
#> 9: (53,62,12), [12,4,10]
#> 10: (53,70,62), [21,9,5]
#> 11: (40,12,83), [12,15,14]
#> 12: (83,12,62), [9,16,11]
#> 13: (12,24,45), [27,8,14]
#> 14: (40,24,12), [13,11,28]
#> 15: (14,40,83), [11,25,24]
#> 16: (21,83,62), [12,17,25]
#> 17: (33,21,62), [16,34,20]
#> 18: (70,77,80), [40,21,33]
#> 19: (8,70,32), [5,22,33]
#> 20: (21,33,14), [26,25,17]
#> 21: (80,62,70), [10,18,29]
#> 22: (57,8,32), [19,6,31]
#> 23: (53,76,57), [38,6,2]
#> 24: (60,40,14), [15,37,36]
#> 25: (14,83,21), [16,20,15]
#> 26: (33,50,14), [37,20,34]
#> 27: (66,45,24), [13,28,0]
#> 28: (40,66,24), [27,14,36]
#> 29: (80,50,62), [34,21,64]
#> 30: (64,57,16), [32,49,31]
#> 31: (64,8,57), [22,30,35]
#> 32: (57,59,16), [69,30,38]
#> 33: (70,8,77), [35,18,19]
#> 34: (50,33,62), [17,29,26]
#> 35: (64,77,8), [33,31,41]
#> 36: (60,66,40), [28,24,53]
#> 37: (60,14,50), [26,43,24]
#> 38: (57,76,59), [52,32,23]
#> 39: (77,36,6), [65,47,48]
#> 40: (51,80,77), [18,47,64]
#> 41: (19,77,64), [35,55,48]
#> 42: (87,36,19), [48,55,61]
#> 43: (71,60,50), [37,45,58]
#> 44: (28,85,50), [45,57,63]
#> 45: (50,85,71), [56,43,44]
#> 46: (51,81,28), [80,57,67]
#> 47: (51,77,6), [39,67,40]
#> 48: (77,19,36), [42,39,41]
#> 49: (64,16,87), [62,55,30]
#> 50: (92,59,27), [51,59,69]
#> 51: (59,5,27), [60,50,52]
#> 52: (5,59,76), [38,0,51]
#> 53: (15,66,60), [36,54,70]
#> 54: (74,15,60), [53,58,71]
#> 55: (87,19,64), [41,49,42]
#> 56: (71,85,74), [74,58,45]
#> 57: (51,28,50), [44,64,46]
#> 58: (74,60,71), [43,56,54]
#> 59: (92,27,99), [60,94,50]
#> 60: (99,27,5), [51,113,59]
#> 61: (35,36,87), [42,78,65]
#> 62: (16,75,87), [68,49,73]
#> 63: (97,85,28), [44,76,74]
#> 64: (51,50,80), [29,40,57]
#> 65: (35,6,36), [39,61,66]
#> 66: (35,81,6), [67,65,77]
#> 67: (51,6,81), [66,46,47]
#> 68: (55,87,75), [62,85,84]
#> 69: (92,16,59), [32,50,73]
#> 70: (100,66,15), [53,71,0]
#> 71: (100,15,74), [54,72,70]
#> 72: (74,42,100), [136,71,75]
#> 73: (16,92,75), [86,62,69]
#> 74: (97,74,85), [56,63,75]
#> 75: (74,97,42), [79,72,74]
#> 76: (28,63,97), [79,63,81]
#> 77: (43,81,35), [66,92,82]
#> 78: (87,7,35), [92,61,84]
#> 79: (63,42,97), [75,76,89]
#> 80: (28,81,65), [91,81,46]
#> 81: (28,65,63), [90,76,80]
#> 82: (89,81,43), [77,104,91]
#> 83: (7,55,78), [98,96,84]
#> 84: (55,7,87), [78,68,83]
#> 85: (75,17,55), [97,68,86]
#> 86: (75,92,17), [87,85,73]
#> 87: (88,17,92), [86,88,102]
#> 88: (92,4,88), [124,87,110]
#> 89: (34,42,63), [79,90,137]
#> 90: (34,63,65), [81,112,89]
#> 91: (89,65,81), [80,82,103]
#> 92: (43,35,7), [78,93,77]
#> 93: (7,22,43), [105,92,106]
#> 94: (92,99,90), [111,110,59]
#> 95: (95,89,37), [104,117,103]
#> 96: (7,78,3), [107,106,83]
#> 97: (93,55,17), [85,114,98]
#> 98: (93,78,55), [83,97,107]
#> 99: (69,17,72), [101,127,114]
#> 100: (96,52,88), [102,124,115]
#> 101: (17,52,72), [123,99,102]
#> 102: (17,88,52), [100,101,87]
#> 103: (95,65,89), [91,95,112]
#> 104: (43,37,89), [95,82,105]
#> 105: (22,37,43), [104,93,119]
#> 106: (7,3,22), [120,93,96]
#> 107: (3,78,93), [98,108,96]
#> 108: (93,69,3), [128,107,114]
#> 109: (4,90,39), [141,135,110]
#> 110: (92,90,4), [109,88,94]
#> 111: (99,31,90), [125,94,113]
#> 112: (95,34,65), [90,103,116]
#> 113: (99,5,31), [148,111,60]
#> 114: (69,93,17), [97,99,108]
#> 115: (84,52,96), [100,131,123]
#> 116: (10,34,95), [112,132,138]
#> 117: (98,95,37), [95,144,132]
#> 118: (47,22,9), [120,145,119]
#> 119: (47,37,22), [105,118,122]
#> 120: (9,22,3), [106,149,118]
#> 121: (68,69,54), [127,153,126]
#> 122: (11,37,47), [119,172,144]
#> 123: (84,72,52), [101,115,130]
#> 124: (96,88,4), [88,135,100]
#> 125: (13,90,31), [111,142,141]
#> 126: (69,68,44), [155,156,121]
#> 127: (54,69,72), [99,130,121]
#> 128: (69,58,3), [146,108,150]
#> 129: (82,54,84), [130,182,153]
#> 130: (84,54,72), [127,123,129]
#> 131: (96,39,84), [161,115,135]
#> 132: (98,10,95), [116,117,134]
#> 133: (94,10,2), [134,173,139]
#> 134: (2,10,98), [132,143,133]
#> 135: (4,39,96), [131,124,109]
#> 136: (61,100,42), [72,137,0]
#> 137: (61,42,34), [89,138,136]
#> 138: (10,61,34), [137,116,139]
#> 139: (94,61,10), [138,133,170]
#> 140: (67,98,11), [144,165,143]
#> 141: (13,39,90), [109,125,147]
#> 142: (31,73,13), [147,125,148]
#> 143: (67,2,98), [134,140,164]
#> 144: (11,98,37), [117,122,140]
#> 145: (79,47,9), [118,151,166]
#> 146: (56,3,58), [128,152,149]
#> 147: (73,39,13), [141,142,176]
#> 148: (31,5,73), [0,142,113]
#> 149: (56,9,3), [120,146,151]
#> 150: (69,30,58), [158,128,156]
#> 151: (79,9,56), [149,167,145]
#> 152: (23,56,58), [146,154,157]
#> 153: (82,68,54), [121,129,155]
#> 154: (58,26,23), [157,152,158]
#> 155: (68,82,44), [159,126,153]
#> 156: (69,44,30), [160,150,126]
#> 157: (26,56,23), [152,154,167]
#> 158: (58,30,26), [169,154,150]
#> 159: (29,44,82), [155,175,160]
#> 160: (29,30,44), [156,159,169]
#> 161: (84,39,46), [176,182,131]
#> 162: (46,41,82), [175,182,174]
#> 163: (25,49,94), [170,173,178]
#> 164: (2,67,25), [171,173,143]
#> 165: (11,20,67), [171,140,172]
#> 166: (79,20,47), [172,145,183]
#> 167: (79,56,26), [157,168,151]
#> 168: (26,38,79), [185,167,187]
#> 169: (29,26,30), [158,160,181]
#> 170: (94,49,61), [177,139,163]
#> 171: (67,20,25), [179,164,165]
#> 172: (11,47,20), [166,165,122]
#> 173: (2,25,94), [163,133,164]
#> 174: (29,41,46), [162,180,175]
#> 175: (41,29,82), [159,162,174]
#> 176: (39,73,46), [0,161,147]
#> 177: (91,61,49), [170,178,0]
#> 178: (91,49,25), [163,179,177]
#> 179: (91,25,20), [171,184,178]
#> 180: (86,29,46), [174,0,181]
#> 181: (26,29,86), [180,187,169]
#> 182: (84,46,82), [162,129,161]
#> 183: (79,48,20), [184,166,185]
#> 184: (48,91,20), [179,183,0]
#> 185: (79,38,48), [186,183,168]
#> 186: (38,86,48), [0,185,187]
#> 187: (26,86,38), [186,168,181]
#> boundary nodes:  45 76 5 73 46 86 48 91 61 100 66 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.121679902 0.035411827 0.122888869 0.079584065 0.097734052 0.082601369 
#>                                                                         
#> 0.117471914 0.114237487 0.097915569 0.089201379 0.101412997 0.049591039 
#>                                                                         
#> 0.070238706 0.039296978 0.042964607 0.071060665 0.037463158 0.064937948 
#>                                                                         
#> 0.261052540 0.106796021 0.028930917 0.098450162 0.076822327 0.116568896 
#>                                                                         
#> 0.069030402 0.126597852 0.136736679 0.088431925 0.046523116 0.089439483 
#>                                                                         
#> 0.291576029 0.098519811 0.096353054 0.044550284 0.105371712 0.097980332 
#>                                                                         
#> 0.036449935 0.105882847 0.090362193 0.207766950 0.062985883 0.047146712 
#>                                                                         
#> 0.181884749 0.092823703 0.053626360 0.092218299 0.150848525 0.106611377 
#>                                                                         
#> 0.030785504 0.098007318 0.048120467 0.097768852 0.081294796 0.062336247 
#>                                                                         
#> 0.213014223 0.083401425 0.029897839 0.270095245 0.103461037 0.023453230 
#>                                                                         
#> 0.122826005 0.048776056 0.129248628 0.133208461 0.126449625 0.082680862 
#>                                                                         
#> 0.100721816 0.133568738 0.259849229 0.063518245 0.118708966 0.058448172 
#>                                                                         
#> 0.222979426 0.035040004 0.041475783 0.125349843 0.106699611 0.059804024 
#>                                                                         
#> 0.115607754 0.100257460 0.181587485 0.062640731 0.126178326 0.171547283 
#>                                                                         
#> 0.113833413 0.027685948 0.159543535 0.256161743 0.116247722 0.141286767 
#>                                                                         
#> 0.128979247 0.141756257 0.078094357 0.124412440 0.076182913 0.113953743 
#>                                                                         
#> 0.126423530 0.182917329 0.190242673 0.281740233 0.163152301 0.115529095 
#>                                                                         
#> 0.055338933 0.066088214 0.089136752 0.083468874 0.155991738 0.196296596 
#>                                                                         
#> 0.147872133 0.075806304 0.009393603 0.041524774 0.155461376 0.090180146 
#>                                                                         
#> 0.151331413 0.416403242 0.171971033 0.063006372 0.185879762 0.104853497 
#>                                                                         
#> 0.011228915 0.007703358 0.106461673 0.036205649 0.091685691 0.076102937 
#>                                                                         
#> 0.591414273 0.081673630 0.127077472 0.160592058 0.082305568 0.211950623 
#>                                                                         
#> 0.092881861 0.141443087 0.161348128 0.234564002 0.137356932 0.092561965 
#>                                                                         
#> 0.102155891 0.745973113 0.040424724 0.108472042 0.226906116 0.173459849 
#>                                                                         
#> 0.399225341 0.321137514 0.043280256 0.149449780 0.142010159 0.050371077 
#>                                                                         
#> 0.146926937 0.105192527 0.082956374 0.105402024 0.087247483 0.135698600 
#>                                                                         
#> 0.055504504 0.067803950 0.108568759 0.042814501 0.163161661 0.045180896 
#>                                                                         
#> 0.022993034 0.138568249 0.163337074 0.133555505 0.113220272 0.091960379 
#>                                                                         
#> 0.138961710 0.175238760 0.090972307 0.100891152 0.075362838 0.098915489 
#>                                                                         
#> 0.031551079 0.106896972 0.117171916 0.127878434 0.334104295 0.115897014 
#>                                                                         
#> 0.079694048 0.140010747 0.137073123 0.040837974 0.104264839 0.138136300 
#>                                                                         
#> 0.086221266 0.053448559 0.117753880 0.076302756 0.106254108 0.166070822 
#>                                                                         
#> 0.067837372 0.056513572 0.206052113 0.086240589 0.063079518 0.174881543 
#>                                                                         
#> 0.066826049 0.139858838 0.081002037 0.118212990 0.098178796 0.080717673 
#>                                                                         
#> 0.099119721 0.071188204 0.147066010 0.179182520 0.183382346 0.147292294 
#>                                                                         
#> 0.098432084 0.149995943 0.165710745 0.149699315 0.191953254 0.093341695 
#>                                                                         
#> 0.176216662 0.104254544 0.135104592 0.219891313 0.157728212 0.054689506 
#>                                                                         
#> 0.096058472 0.156994858 0.145949985 0.145699680 0.153864946 0.167361622 
#>                                                                         
#> 0.313938564 0.181939757 0.130195444 0.074982625 0.330644511 0.165107174 
#>                                                                         
#> 0.090987853 0.073959715 0.154855027 0.129921634 0.171269087 0.204389740 
#>                                                                         
#> 0.160883959 0.194047534 0.175125363 0.080743867 0.022165415 0.149755128 
#>                                                                         
#> 0.175129425 0.028655724 0.064294983 0.482153097 0.157141538 0.259008244 
#>                                                                         
#> 0.155843360 0.075931624 0.088108537 0.126706517 0.053903024 0.087499475 
#>                                                                         
#> 0.170687674 0.230088796 0.308290771 0.177247593 0.209251536 0.157868336 
#>                                                                         
#> 0.177105724 0.011250667 0.062723130 0.096980377 0.129319384 0.084580790 
#>                                                                         
#> 0.074279029 0.185296604 0.280803563 0.090140413 0.051839457 0.222730618 
#>                                                                         
#> 0.074892862 0.077218728 0.360536985 0.154485486 0.159061415 0.142983091 
#>                                                 
#> 0.137656763 0.051531559 0.174824980 0.719360257 
#> 
#> $alpha
#> [1] 0.2781132
#> 
#> attr(,"class")
#> [1] "scree"