Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1          y1          x2          y2         mx1
#>    67   37 0.1870395163 0.968840270 0.122744468 0.921630979  0.18519675
#>    37   17 0.1227444685 0.921630979 0.205718272 0.799959867  0.16608761
#>    17   70 0.2057182720 0.799959867 0.323403437 0.731309665  0.28622936
#>    70   26 0.3234034372 0.731309665 0.407699236 0.847323419  0.31922323
#>    26   15 0.4076992357 0.847323419 0.309711113 0.945928442  0.32414610
#>    15   67 0.3097111134 0.945928442 0.187039516 0.968840270  0.24414617
#>    59   47 0.0943717650 0.705953209 0.065642951 0.656470509  0.08115474
#>    47   23 0.0656429512 0.656470509 0.174318122 0.647460371  0.11965717
#>    23   73 0.1743181222 0.647460371 0.145452588 0.702735307  0.12672955
#>    73   59 0.1454525876 0.702735307 0.094371765 0.705953209  0.12174595
#>    47   85 0.0656429512 0.656470509 0.066158378 0.516289714 -0.11697863
#>    85   80 0.0661583775 0.516289714 0.081857743 0.512047739  0.07744122
#>    80   19 0.0818577427 0.512047739 0.093753993 0.526214514  0.07830972
#>    19   23 0.0937539928 0.526214514 0.174318122 0.647460371  0.12855494
#>    87  100 0.1503327543 0.033292910 0.164447894 0.200834633  0.09798084
#>   100   40 0.1644478942 0.200834633 0.000000000 0.172395956  0.09232273
#>    70   71 0.3234034372 0.731309665 0.411979792 0.702673390  0.37585642
#>    71   84 0.4119797922 0.702673390 0.690650887 0.753970156  0.55258083
#>    84   42 0.6906508865 0.753970156 0.584718989 0.907511407  0.55618724
#>    42   26 0.5847189891 0.907511407 0.407699236 0.847323419  0.51354855
#>    40   79 0.0000000000 0.172395956 0.129460092 0.240906188  0.06669542
#>    79   62 0.1294600918 0.240906188 0.007047782 0.287171795  0.06401084
#>    62   72 0.0070477820 0.287171795 0.004627099 0.862872955 -2.36896321
#>    79   36 0.1294600918 0.240906188 0.188916871 0.288252286  0.12714460
#>    36   53 0.1889168710 0.288252286 0.239656781 0.323227575  0.19998186
#>    53   65 0.2396567806 0.323227575 0.172281896 0.435993917  0.22763331
#>    65   13 0.1722818956 0.435993917 0.040585428 0.429518677  0.10823663
#>    13   62 0.0405854275 0.429518677 0.007047782 0.287171795  0.07810493
#>    74   95 0.9717863502 0.925294758 0.724062652 0.874696574  0.86194570
#>    95   12 0.7240626520 0.874696574 0.795655336 0.680854684  0.85417958
#>    12   41 0.7956553358 0.680854684 0.855861689 0.608524306  0.89986096
#>    41   69 0.8558616887 0.608524306 0.919945949 0.519523319  0.98446481
#>    69   82 0.9199459485 0.519523319 0.975662167 0.485608867  1.02114308
#>    82   20 0.9756621670 0.485608867 0.979266974 0.424368998  2.51851103
#>    20   58 0.9792669739 0.424368998 0.994807120 0.269509899  2.85973429
#>    58    9 0.9948071196 0.269509899 1.000000000 0.220212154  3.83961266
#>    58   92 0.9948071196 0.269509899 0.965667546 0.249083632  0.99084107
#>    92   50 0.9656675457 0.249083632 0.870687059 0.213077199  0.93733544
#>    50   16 0.8706870592 0.213077199 0.890319132 0.009455548  0.94083691
#>    21   38 0.6694818991 0.347225637 0.521843794 0.245766664  0.59879140
#>    38   93 0.5218437942 0.245766664 0.534119601 0.197432756  0.58559732
#>    93   44 0.5341196009 0.197432756 0.603935669 0.117470887  0.59492769
#>    44   30 0.6039356686 0.117470887 0.650666729 0.084015930  0.66291959
#>    30   98 0.6506667293 0.084015930 0.745569261 0.180159755  0.67899875
#>    98   21 0.7455692613 0.180159755 0.669481899 0.347225637  0.72261808
#>    14   75 0.5578998374 0.439839542 0.486609253 0.477634283  0.52971316
#>    75   76 0.4866092529 0.477634283 0.439749537 0.458456613  0.48446558
#>    76   33 0.4397495373 0.458456613 0.547509744 0.416461632  0.48292421
#>    33   14 0.5475097436 0.416461632 0.557899837 0.439839542  0.55632237
#>    95   48 0.7240626520 0.874696574 0.700557366 0.758794839  0.81111747
#>    48   84 0.7005573659 0.758794839 0.690650887 0.753970156  0.72634185
#>    84   68 0.6906508865 0.753970156 0.774539745 0.623779716  0.71570328
#>    68   57 0.7745397451 0.623779716 0.831606945 0.609214194  0.80987614
#>    57   41 0.8316069445 0.609214194 0.855861689 0.608524306  0.84521379
#>    85   52 0.0661583775 0.516289714 0.046520042 0.721614686 -0.32003516
#>    52   55 0.0465200416 0.721614686 0.027115575 0.776207895 -0.42414189
#>    55   62 0.0271155754 0.776207895 0.007047782 0.287171795 -0.86567584
#>    65   51 0.1722818956 0.435993917 0.075883407 0.468032982  0.10788329
#>    51   85 0.0758834074 0.468032982 0.066158378 0.516289714  0.02173345
#>    33   94 0.5475097436 0.416461632 0.646066046 0.449902757  0.60662677
#>    94   90 0.6460660465 0.449902757 0.626110144 0.480393071  0.60220014
#>    90   88 0.6261101441 0.480393071 0.607943349 0.524519538  0.57670247
#>    88   75 0.6079433495 0.524519538 0.486609253 0.477634283  0.54619063
#>    49   83 0.4080477800 0.977311861 0.194333594 0.997979134  0.31267467
#>    83   67 0.1943335944 0.997979134 0.187039516 0.968840270  0.25044238
#>    26   49 0.4076992357 0.847323419 0.408047780 0.977311861  0.37458549
#>    50   98 0.8706870592 0.213077199 0.745569261 0.180159755  0.83181851
#>    98   22 0.7455692613 0.180159755 0.671417216 0.043974941  0.78754038
#>    22   97 0.6714172162 0.043974941 0.426592052 0.000000000  0.67207466
#>    37   56 0.1227444685 0.921630979 0.025623067 0.820053103  0.11780481
#>    56   77 0.0256230671 0.820053103 0.051390784 0.738775688  0.11496183
#>    77   59 0.0513907837 0.738775688 0.094371765 0.705953209  0.12179519
#>    73   70 0.1454525876 0.702735307 0.323403437 0.731309665  0.23481125
#>    92   20 0.9656675457 0.249083632 0.979266974 0.424368998  0.91234755
#>    20   61 0.9792669739 0.424368998 0.878397155 0.470937787  0.90609772
#>    61   91 0.8783971552 0.470937787 0.788521291 0.505392679  0.82376397
#>    91   24 0.7885212911 0.505392679 0.754326816 0.513998775  0.75214847
#>    24   94 0.7543268155 0.513998775 0.646066046 0.449902757  0.73589178
#>    94   21 0.6460660465 0.449902757 0.669481899 0.347225637  0.73461272
#>    80   51 0.0818577427 0.512047739 0.075883407 0.468032982  0.13085111
#>    65   23 0.1722818956 0.435993917 0.174318122 0.647460371  0.20234621
#>    82   61 0.9756621670 0.485608867 0.878397155 0.470937787  0.93095474
#>    33   45 0.5475097436 0.416461632 0.527651660 0.354188107  0.59930002
#>    45   38 0.5276516604 0.354188107 0.521843794 0.245766664  0.59588938
#>    30   54 0.6506667293 0.084015930 0.485156827 0.078256923  0.57014723
#>    54   97 0.4851568267 0.078256923 0.426592052 0.000000000  0.55991027
#>    65   96 0.1722818956 0.435993917 0.401941581 0.518918745  0.26422114
#>    96   29 0.4019415805 0.518918745 0.361351774 0.674485659  0.28091922
#>    29   70 0.3613517739 0.674485659 0.323403437 0.731309665  0.26887028
#>    91   99 0.7885212911 0.505392679 0.841177713 0.553126120  0.78471448
#>    99   68 0.8411777127 0.553126120 0.774539745 0.623779716  0.78341133
#>    68   88 0.7745397451 0.623779716 0.607943349 0.524519538  0.68574336
#>    86   53 0.3086837246 0.361022733 0.239656781 0.323227575  0.31134126
#>    53   46 0.2396567806 0.323227575 0.279887486 0.178175370  0.29711497
#>    46   43 0.2798874858 0.178175370 0.348964278 0.141624779  0.32154930
#>    43   31 0.3489642778 0.141624779 0.384564234 0.088345202  0.43683323
#>    31   54 0.3845642340 0.088345202 0.485156827 0.078256923  0.44289738
#>    54   93 0.4851568267 0.078256923 0.534119601 0.197432756  0.44412257
#>    93   81 0.5341196009 0.197432756 0.482094345 0.265467497  0.44428133
#>    81   34 0.4820943447 0.265467497 0.425653895 0.339235045  0.40238717
#>    34   86 0.4256538952 0.339235045 0.308683725 0.361022733  0.35536307
#>    42   39 0.5847189891 0.907511407 0.522681054 0.975693076  0.49312568
#>    39   49 0.5226810541 0.975693076 0.408047780 0.977311861  0.46445589
#>    77   52 0.0513907837 0.738775688 0.046520042 0.721614686  0.06147748
#>    52   47 0.0465200416 0.721614686 0.065642951 0.656470509 -0.17957368
#>    89   87 0.3114272231 0.067572913 0.150332754 0.033292910  0.23547978
#>    87   97 0.1503327543 0.033292910 0.426592052 0.000000000  0.26325166
#>    97   31 0.4265920520 0.000000000 0.384564234 0.088345202  0.37236973
#>    31   89 0.3845642340 0.088345202 0.311427223 0.067572913  0.34883245
#>    96   64 0.4019415805 0.518918745 0.447541519 0.544171708  0.38777283
#>    64   71 0.4475415194 0.544171708 0.411979792 0.702673390  0.42369978
#>    44   54 0.6039356686 0.117470887 0.485156827 0.078256923  0.57001391
#>    43   89 0.3489642778 0.141624779 0.311427223 0.067572913  0.34215067
#>    95   78 0.7240626520 0.874696574 0.843623833 1.000000000  0.80930455
#>    78   39 0.8436238327 1.000000000 0.522681054 0.975693076  0.62440036
#>    39   60 0.5226810541 0.975693076 0.684147726 0.952188535  0.63058119
#>    60   95 0.6841477264 0.952188535 0.724062652 0.874696574  0.73006525
#>    76   45 0.4397495373 0.458456613 0.527651660 0.354188107  0.48041643
#>    76   86 0.4397495373 0.458456613 0.308683725 0.361022733  0.37745974
#>    81   45 0.4820943447 0.265467497 0.527651660 0.354188107  0.48020287
#>    72   66 0.0046270986 0.862872955 0.080219530 0.937277715  0.03207358
#>    66   63 0.0802195305 0.937277715 0.107717101 0.960119183  0.07124890
#>    63   83 0.1077171012 0.960119183 0.194333594 0.997979134  0.08293963
#>    79  100 0.1294600918 0.240906188 0.164447894 0.200834633  0.17470171
#>   100   46 0.1644478942 0.200834633 0.279887486 0.178175370  0.23003945
#>    66   56 0.0802195305 0.937277715 0.025623067 0.820053103  0.07842424
#>    67   63 0.1870395163 0.968840270 0.107717101 0.960119183  0.14854350
#>    78   83 0.8436238327 1.000000000 0.194333594 0.997979134  0.51197890
#>    42   60 0.5847189891 0.907511407 0.684147726 0.952188535  0.60725621
#>    57   99 0.8316069445 0.609214194 0.841177713 0.553126120  0.84297843
#>    99   69 0.8411777127 0.553126120 0.919945949 0.519523319  0.89436219
#>    84   95 0.6906508865 0.753970156 0.724062652 0.874696574  0.64065871
#>    46   89 0.2798874858 0.178175370 0.311427223 0.067572913  0.29468811
#>    77   55 0.0513907837 0.738775688 0.027115575 0.776207895  0.01292917
#>    86   65 0.3086837246 0.361022733 0.172281896 0.435993917  0.24189314
#>    56   72 0.0256230671 0.820053103 0.004627099 0.862872955 -0.08068039
#>    88   71 0.6079433495 0.524519538 0.411979792 0.702673390  0.54173397
#>    96   76 0.4019415805 0.518918745 0.439749537 0.458456613  0.44136870
#>    86   96 0.3086837246 0.361022733 0.401941581 0.518918745  0.28582859
#>    84   88 0.6906508865 0.753970156 0.607943349 0.524519538  0.65332127
#>     1   37 0.2394844432 0.868392250 0.122744468 0.921630979  0.18519675
#>     1   17 0.2394844432 0.868392250 0.205718272 0.799959867  0.16608761
#>     1   70 0.2394844432 0.868392250 0.323403437 0.731309665  0.28622936
#>     1   26 0.2394844432 0.868392250 0.407699236 0.847323419  0.31922323
#>     1   15 0.2394844432 0.868392250 0.309711113 0.945928442  0.32414610
#>     1   67 0.2394844432 0.868392250 0.187039516 0.968840270  0.24414617
#>     2   47 0.1022727633 0.699891507 0.065642951 0.656470509  0.08115474
#>     2   23 0.1022727633 0.699891507 0.174318122 0.647460371  0.11965717
#>     2   73 0.1022727633 0.699891507 0.145452588 0.702735307  0.12672955
#>     2   59 0.1022727633 0.699891507 0.094371765 0.705953209  0.12174595
#>     3   85 0.0739201886 0.541939437 0.066158378 0.516289714 -0.11697863
#>     3   80 0.0739201886 0.541939437 0.081857743 0.512047739  0.07744122
#>     3   19 0.0739201886 0.541939437 0.093753993 0.526214514  0.07830972
#>     3   23 0.0739201886 0.541939437 0.174318122 0.647460371  0.12855494
#>     3   47 0.0739201886 0.541939437 0.065642951 0.656470509  0.11588246
#>     4  100 0.0139627554 0.062379183 0.164447894 0.200834633  0.09798084
#>     4   40 0.0139627554 0.062379183 0.000000000 0.172395956  0.09232273
#>     5   71 0.4201805218 0.772351362 0.411979792 0.702673390  0.37585642
#>     5   84 0.4201805218 0.772351362 0.690650887 0.753970156  0.55258083
#>     5   42 0.4201805218 0.772351362 0.584718989 0.907511407  0.55618724
#>     5   26 0.4201805218 0.772351362 0.407699236 0.847323419  0.51354855
#>     5   70 0.4201805218 0.772351362 0.323403437 0.731309665  0.35159385
#>     6   79 0.0002177514 0.233949568 0.129460092 0.240906188  0.06669542
#>     6   62 0.0002177514 0.233949568 0.007047782 0.287171795  0.06401084
#>     6   72 0.0002177514 0.233949568 0.004627099 0.862872955 -2.36896321
#>     7   36 0.1698106904 0.352463049 0.188916871 0.288252286  0.12714460
#>     7   53 0.1698106904 0.352463049 0.239656781 0.323227575  0.19998186
#>     7   65 0.1698106904 0.352463049 0.172281896 0.435993917  0.22763331
#>     7   13 0.1698106904 0.352463049 0.040585428 0.429518677  0.10823663
#>     7   62 0.1698106904 0.352463049 0.007047782 0.287171795  0.07810493
#>     7   79 0.1698106904 0.352463049 0.129460092 0.240906188  0.08889577
#>     8   95 0.9900662483 0.764447526 0.724062652 0.874696574  0.86194570
#>     8   12 0.9900662483 0.764447526 0.795655336 0.680854684  0.85417958
#>     8   41 0.9900662483 0.764447526 0.855861689 0.608524306  0.89986096
#>     8   69 0.9900662483 0.764447526 0.919945949 0.519523319  0.98446481
#>     8   82 0.9900662483 0.764447526 0.975662167 0.485608867  1.02114308
#>     8   20 0.9900662483 0.764447526 0.979266974 0.424368998  2.51851103
#>     8   58 0.9900662483 0.764447526 0.994807120 0.269509899  2.85973429
#>     8    9 0.9900662483 0.764447526 1.000000000 0.220212154  3.83961266
#>     9   92 1.0000000000 0.220212154 0.965667546 0.249083632  0.99084107
#>     9   50 1.0000000000 0.220212154 0.870687059 0.213077199  0.93733544
#>     9   16 1.0000000000 0.220212154 0.890319132 0.009455548  0.94083691
#>    10   38 0.6468548765 0.216160065 0.521843794 0.245766664  0.59879140
#>    10   93 0.6468548765 0.216160065 0.534119601 0.197432756  0.58559732
#>    10   44 0.6468548765 0.216160065 0.603935669 0.117470887  0.59492769
#>    10   30 0.6468548765 0.216160065 0.650666729 0.084015930  0.66291959
#>    10   98 0.6468548765 0.216160065 0.745569261 0.180159755  0.67899875
#>    10   21 0.6468548765 0.216160065 0.669481899 0.347225637  0.72261808
#>    11   75 0.5440356441 0.431865236 0.486609253 0.477634283  0.52971316
#>    11   76 0.5440356441 0.431865236 0.439749537 0.458456613  0.48446558
#>    11   33 0.5440356441 0.431865236 0.547509744 0.416461632  0.48292421
#>    11   14 0.5440356441 0.431865236 0.557899837 0.439839542  0.55632237
#>    12   48 0.7956553358 0.680854684 0.700557366 0.758794839  0.81111747
#>    12   84 0.7956553358 0.680854684 0.690650887 0.753970156  0.72634185
#>    12   68 0.7956553358 0.680854684 0.774539745 0.623779716  0.71570328
#>    12   57 0.7956553358 0.680854684 0.831606945 0.609214194  0.80987614
#>    13   52 0.0405854275 0.429518677 0.046520042 0.721614686 -0.32003516
#>    13   55 0.0405854275 0.429518677 0.027115575 0.776207895 -0.42414189
#>    13   51 0.0405854275 0.429518677 0.075883407 0.468032982  0.10788329
#>    13   85 0.0405854275 0.429518677 0.066158378 0.516289714  0.02173345
#>    14   94 0.5578998374 0.439839542 0.646066046 0.449902757  0.60662677
#>    14   90 0.5578998374 0.439839542 0.626110144 0.480393071  0.60220014
#>    14   88 0.5578998374 0.439839542 0.607943349 0.524519538  0.57670247
#>    15   83 0.3097111134 0.945928442 0.194333594 0.997979134  0.31267467
#>    15   49 0.3097111134 0.945928442 0.408047780 0.977311861  0.37458549
#>    16   98 0.8903191318 0.009455548 0.745569261 0.180159755  0.83181851
#>    16   22 0.8903191318 0.009455548 0.671417216 0.043974941  0.78754038
#>    16   97 0.8903191318 0.009455548 0.426592052 0.000000000  0.67207466
#>    17   56 0.2057182720 0.799959867 0.025623067 0.820053103  0.11780481
#>    17   77 0.2057182720 0.799959867 0.051390784 0.738775688  0.11496183
#>    17   59 0.2057182720 0.799959867 0.094371765 0.705953209  0.12179519
#>    17   73 0.2057182720 0.799959867 0.145452588 0.702735307  0.12485421
#>    18   20 0.8290425672 0.407903004 0.979266974 0.424368998  0.91234755
#>    18   61 0.8290425672 0.407903004 0.878397155 0.470937787  0.90609772
#>    18   91 0.8290425672 0.407903004 0.788521291 0.505392679  0.82376397
#>    18   24 0.8290425672 0.407903004 0.754326816 0.513998775  0.75214847
#>    18   94 0.8290425672 0.407903004 0.646066046 0.449902757  0.73589178
#>    18   21 0.8290425672 0.407903004 0.669481899 0.347225637  0.73461272
#>    18   98 0.8290425672 0.407903004 0.745569261 0.180159755  0.78001121
#>    18   50 0.8290425672 0.407903004 0.870687059 0.213077199  0.78198671
#>    18   92 0.8290425672 0.407903004 0.965667546 0.249083632  0.88520959
#>    19   51 0.0937539928 0.526214514 0.075883407 0.468032982  0.13085111
#>    19   65 0.0937539928 0.526214514 0.172281896 0.435993917  0.13404977
#>    21   33 0.6694818991 0.347225637 0.547509744 0.416461632  0.61207068
#>    21   45 0.6694818991 0.347225637 0.527651660 0.354188107  0.59930002
#>    22   30 0.6714172162 0.043974941 0.650666729 0.084015930  0.73058846
#>    22   54 0.6714172162 0.043974941 0.485156827 0.078256923  0.57014723
#>    23   96 0.1743181222 0.647460371 0.401941581 0.518918745  0.26422114
#>    23   29 0.1743181222 0.647460371 0.361351774 0.674485659  0.28091922
#>    23   70 0.1743181222 0.647460371 0.323403437 0.731309665  0.26887028
#>    24   99 0.7543268155 0.513998775 0.841177713 0.553126120  0.78471448
#>    24   68 0.7543268155 0.513998775 0.774539745 0.623779716  0.78341133
#>    24   88 0.7543268155 0.513998775 0.607943349 0.524519538  0.68574336
#>    24   90 0.7543268155 0.513998775 0.626110144 0.480393071  0.68184538
#>    25   53 0.3547198093 0.199025547 0.239656781 0.323227575  0.31134126
#>    25   46 0.3547198093 0.199025547 0.279887486 0.178175370  0.29711497
#>    25   43 0.3547198093 0.199025547 0.348964278 0.141624779  0.32154930
#>    25   31 0.3547198093 0.199025547 0.384564234 0.088345202  0.43683323
#>    25   54 0.3547198093 0.199025547 0.485156827 0.078256923  0.44289738
#>    25   93 0.3547198093 0.199025547 0.534119601 0.197432756  0.44412257
#>    25   81 0.3547198093 0.199025547 0.482094345 0.265467497  0.44428133
#>    25   34 0.3547198093 0.199025547 0.425653895 0.339235045  0.40238717
#>    25   86 0.3547198093 0.199025547 0.308683725 0.361022733  0.35536307
#>    26   39 0.4076992357 0.847323419 0.522681054 0.975693076  0.49312568
#>    27   77 0.0477042865 0.718941167 0.051390784 0.738775688  0.07432381
#>    27   52 0.0477042865 0.718941167 0.046520042 0.721614686  0.06147748
#>    27   47 0.0477042865 0.718941167 0.065642951 0.656470509 -0.17957368
#>    27   59 0.0477042865 0.718941167 0.094371765 0.705953209  0.06480189
#>    28   87 0.3168822633 0.054187774 0.150332754 0.033292910  0.23547978
#>    28   97 0.3168822633 0.054187774 0.426592052 0.000000000  0.26325166
#>    28   31 0.3168822633 0.054187774 0.384564234 0.088345202  0.37236973
#>    28   89 0.3168822633 0.054187774 0.311427223 0.067572913  0.34883245
#>    29   64 0.3613517739 0.674485659 0.447541519 0.544171708  0.38777283
#>    29   71 0.3613517739 0.674485659 0.411979792 0.702673390  0.42369978
#>    32   78 0.7419853025 0.977904195 0.843623833 1.000000000  0.80930455
#>    32   39 0.7419853025 0.977904195 0.522681054 0.975693076  0.62440036
#>    32   60 0.7419853025 0.977904195 0.684147726 0.952188535  0.63058119
#>    32   95 0.7419853025 0.977904195 0.724062652 0.874696574  0.73006525
#>    34   45 0.4256538952 0.339235045 0.527651660 0.354188107  0.48020287
#>    34   76 0.4256538952 0.339235045 0.439749537 0.458456613  0.46965022
#>    35   66 0.0328407031 0.965632656 0.080219530 0.937277715  0.03207358
#>    35   63 0.0328407031 0.965632656 0.107717101 0.960119183  0.07124890
#>    35   83 0.0328407031 0.965632656 0.194333594 0.997979134  0.08293963
#>    36  100 0.1889168710 0.288252286 0.164447894 0.200834633  0.17470171
#>    36   46 0.1889168710 0.288252286 0.279887486 0.178175370  0.23003945
#>    37   63 0.1227444685 0.921630979 0.107717101 0.960119183  0.14854350
#>    37   66 0.1227444685 0.921630979 0.080219530 0.937277715  0.10408263
#>    38   81 0.5218437942 0.245766664 0.482094345 0.265467497  0.52397566
#>    39   83 0.5226810541 0.975693076 0.194333594 0.997979134  0.51197890
#>    41   99 0.8558616887 0.608524306 0.841177713 0.553126120  0.84297843
#>    42   95 0.5847189891 0.907511407 0.724062652 0.874696574  0.64065871
#>    46   87 0.2798874858 0.178175370 0.150332754 0.033292910  0.20712596
#>    55   56 0.0271155754 0.776207895 0.025623067 0.820053103  0.10610302
#>    55   72 0.0271155754 0.776207895 0.004627099 0.862872955 -0.08068039
#>    61   69 0.8783971552 0.470937787 0.919945949 0.519523319  0.92821870
#>    61   99 0.8783971552 0.470937787 0.841177713 0.553126120  0.87269136
#>    64   76 0.4475415194 0.544171708 0.439749537 0.458456613  0.44136870
#>    64   75 0.4475415194 0.544171708 0.486609253 0.477634283  0.44979263
#>    64   88 0.4475415194 0.544171708 0.607943349 0.524519538  0.52935015
#>     4   87 0.0139627554 0.062379183 0.150332754 0.033292910  0.09798084
#>     6   40 0.0002177514 0.233949568 0.000000000 0.172395956  0.06669542
#>     8   74 0.9900662483 0.764447526 0.971786350 0.925294758  0.86194570
#>    35   72 0.0328407031 0.965632656 0.004627099 0.862872955  0.03207358
#>    74   78 0.9717863502 0.925294758 0.843623833 1.000000000  0.85396043
#>          my1          mx2         my2 bp1 bp2
#>   0.90396312   0.14854350  0.95388173   0   0
#>   0.86206129   0.11780481  0.82913476   0   0
#>   0.80278049   0.23481125  0.71463583   0   0
#>   0.82297862   0.35159385  0.79945808   0   0
#>   0.86228307   0.37458549  0.91240690   0   0
#>   0.93474121   0.25044238  0.96845155   0   0
#>   0.68054571   0.06480189  0.69003990   0   0
#>   0.64806518   0.11588246  0.60253671   0   0
#>   0.65778330   0.23487205  0.71425720   0   0
#>   0.73345347   0.12485421  0.78279388   0   0
#>   0.58570769  -0.29820964  0.58504133   0   0
#>   0.52687469   0.06789124  0.49153064   0   0
#>   0.52710532   0.13085111  0.48298479   0   0
#>   0.59047948   0.20234621  0.54144746   0   0
#>   0.12206893   0.20712596  0.11287362   0   0
#>   0.12821865   0.08556889  0.16727304   0   0
#>   0.74224650   0.36885079  0.72057705   0   0
#>   0.72144697   0.56191427  0.67074278   0   0
#>   0.77451351   0.64065871  0.83279246   0   0
#>   0.82642016   0.49312568  0.88648615   0   0
#>   0.20293721   0.08556889  0.16727304   0   0
#>   0.25281234   0.08889577  0.31865440   0   0
#>   0.56503692  -0.94179666  0.57103780   0   0
#>   0.30481963   0.17470171  0.24509787   0   0
#>   0.32649266   0.25343640  0.24894424   0   0
#>   0.39255439   0.24189314  0.40107426   0   0
#>   0.39608667   0.10788329  0.40327313   0   0
#>   0.34555463  -0.86567584  0.56791416   0   0
#>   0.83134928   0.85396043  0.87044434   0   0
#>   0.81261156   0.81111747  0.79670720   0   0
#>   0.70637089   0.84521379  0.66088367   0   0
#>   0.63355157   0.89436219  0.56867409   0   0
#>   0.62305081   0.92821870  0.47039036   0   0
#>   0.54570067   0.93095474  0.45225119   0   0
#>   0.53486504   0.92328601  0.34054202   0   0
#>   0.54425103   0.99084107  0.24416975   0   0
#>   0.24416975   0.92328601  0.34054202   0   0
#>   0.18054364   0.88520959  0.31804513   0   0
#>   0.11708343   0.83181851  0.10657248   0   0
#>   0.29194363   0.59588938  0.29616650   0   0
#>   0.23623288   0.47893531  0.20914295   0   0
#>   0.18006560   0.53475339  0.12752640   0   0
#>   0.15049642   0.57001391  0.02072251   0   0
#>   0.15096025   0.73058846  0.10003659   0   0
#>   0.27056632   0.78001121  0.29670506   0   0
#>   0.47280578   0.54619063  0.50388651   0   0
#>   0.41603367   0.44979263  0.50075535   0   0
#>   0.40998871   0.48041643  0.40355369   0   0
#>   0.42654279   0.60662677  0.40418547   0   0
#>   0.79670720   0.66115309  0.82712051   0   0
#>   0.69326900   0.66115309  0.82712051   0   0
#>   0.67799047   0.65332127  0.63779431   0   0
#>   0.64315005   0.79212893  0.57361713   0   0
#>   0.66088367   0.84297843  0.58229400   0   0
#>   0.58295382  -0.29820964  0.58504133   0   0
#>   0.58506899   0.01292917  0.74042038   0   0
#>   0.56791416  -0.94179666  0.57103780   0   0
#>   0.40327313   0.13404977  0.48200232   0   0
#>   0.48222860   0.06789124  0.49153064   0   0
#>   0.40418547   0.61207068  0.38814141   0   0
#>   0.44296825   0.68946223  0.50008127   0   0
#>   0.48585487   0.68184538  0.52914203   0   0
#>   0.50388651   0.52935015  0.54746789   0   0
#>   1.10639799   0.49362159  2.97751682   0   0
#>   0.96845155   0.14401252  0.99509321   0   0
#>   0.91240690   0.46445589  0.91216592   0   0
#>   0.10657248   0.78198671  0.29598099   0   0
#>   0.06902652   0.73058846  0.10003659   0   0
#>  -0.66318998   0.55991027 -0.03872836   0   0
#>   0.82913476   0.07842424  0.86678761   0   0
#>   0.80365321   0.10610302  0.80084467   0   0
#>   0.78641714   0.07432381  0.72425344   0   0
#>   0.71463583   0.23487205  0.71425720   0   0
#>   0.34139067   0.92328601  0.34054202   0   0
#>   0.39840990   0.93095474  0.45225119   0   0
#>   0.46287508   0.83899993  0.50261822   0   0
#>   0.43310832   0.78471448  0.56250239   0   0
#>   0.42165988   0.68946223  0.50008127   0   0
#>   0.41608752   0.61207068  0.38814141   0   0
#>   0.48298479   0.06789124  0.49153064   0   0
#>   0.54144746   0.26422114  0.54085166   0   0
#>   0.45225119   0.92821870  0.47039036   0   0
#>   0.36564351   0.48041643  0.40355369   0   0
#>   0.29616650   0.52397566  0.30001874   0   0
#>   0.01689095   0.57001391  0.02072251   0   0
#>  -0.03872836   0.43220684  0.05684048   0   0
#>   0.54085166   0.28582859  0.48100999   0   0
#>   0.57042085   0.38777283  0.59830061   0   0
#>   0.65380783   0.36885079  0.72057705   0   0
#>   0.56250239   0.83899993  0.50261822   0   0
#>   0.56539500   0.79212893  0.57361713   0   0
#>   0.58337768   0.65332127  0.63779431   0   0
#>   0.27423813   0.24189314  0.40107426   0   0
#>   0.26105863   0.25343640  0.24894424   0   0
#>   0.17336259   0.29468811  0.12259775   0   0
#>   0.16180316   0.34215067  0.09853887   0   0
#>   0.16343833   0.43220684  0.05684048   0   0
#>   0.16476160   0.53475339  0.12752640   0   0
#>   0.18264351   0.47893531  0.20914295   0   0
#>   0.26295798   0.48020287  0.32249572   0   0
#>   0.28674817   0.37745974  0.40537712   0   0
#>   0.88648615   0.60725621  0.99033258   0   0
#>   0.91216592   0.49362159  2.97751682   0   0
#>   0.72664110   0.01292917  0.74042038   0   0
#>   0.61986661  -0.29820964  0.58504133   0   0
#>   0.02881679   0.22005268  0.10131448   0   0
#>  -0.19254830  -0.16673587 -3.76051677   0   1
#>   0.02837460   0.43220684  0.05684048   0   0
#>   0.07501305   0.34215067  0.09853887   0   0
#>   0.59830061   0.44136870  0.50152114   0   0
#>   0.62206272   0.54173397  0.64854509   0   0
#>   0.02072251   0.53475339  0.12752640   0   0
#>   0.09853887   0.29468811  0.12259775   0   0
#>   0.91305378   0.85396043  0.87044434   0   0
#>   1.76359485   0.51197890  3.24798081   0   0
#>   1.15056568   0.60725621  0.99033258   0   0
#>   0.92681419   0.65351483  0.88738424   0   0
#>   0.40355369   0.46965022  0.39447738   0   0
#>   0.40537712   0.34894876  0.44372944   0   0
#>   0.32249572   0.52397566  0.30001874   0   0
#>   0.91059028   0.07244871  0.86957067   0   0
#>   0.97604909   0.10408263  0.93652239   0   0
#>   1.13481655   0.14401252  0.99509321   0   0
#>   0.24509787   0.08556889  0.16727304   0   0
#>   0.22960834   0.20712596  0.11287362   0   0
#>   0.86678761   0.07244871  0.86957067   0   0
#>   0.95388173   0.14401252  0.99509321   0   0
#>   3.24798081   0.50718537  4.78810918   0   1
#>   0.99033258   0.65351483  0.88738424   0   0
#>   0.58229400   0.79212893  0.57361713   0   0
#>   0.56867409   0.87269136  0.51787557   0   0
#>   0.83279246   0.66115309  0.82712051   0   0
#>   0.12259775   0.22005268  0.10131448   0   0
#>   0.74042038   0.10610302  0.80084467   0   0
#>   0.40107426   0.28582859  0.48100999   0   0
#>   0.79448648   0.07244871  0.86957067   0   0
#>   0.64854509   0.56191427  0.67074278   0   0
#>   0.50152114   0.34894876  0.44372944   0   0
#>   0.48100999   0.34894876  0.44372944   0   0
#>   0.63779431   0.56191427  0.67074278   0   0
#>   0.90396312   0.16608761  0.86206129   0   0
#>   0.86206129   0.28622936  0.80278049   0   0
#>   0.80278049   0.31922323  0.82297862   0   0
#>   0.82297862   0.32414610  0.86228307   0   0
#>   0.86228307   0.24414617  0.93474121   0   0
#>   0.93474121   0.18519675  0.90396312   0   0
#>   0.68054571   0.11965717  0.64806518   0   0
#>   0.64806518   0.12672955  0.65778330   0   0
#>   0.65778330   0.12174595  0.73345347   0   0
#>   0.73345347   0.08115474  0.68054571   0   0
#>   0.58570769   0.07744122  0.52687469   0   0
#>   0.52687469   0.07830972  0.52710532   0   0
#>   0.52710532   0.12855494  0.59047948   0   0
#>   0.59047948   0.11588246  0.60253671   0   0
#>   0.60253671  -0.11697863  0.58570769   0   0
#>   0.12206893   0.09232273  0.12821865   0   0
#>   0.12821865  -8.85311299 -1.00708942   0   1
#>   0.74224650   0.55258083  0.72144697   0   0
#>   0.72144697   0.55618724  0.77451351   0   0
#>   0.77451351   0.51354855  0.82642016   0   0
#>   0.82642016   0.35159385  0.79945808   0   0
#>   0.79945808   0.37585642  0.74224650   0   0
#>   0.20293721   0.06401084  0.25281234   0   0
#>   0.25281234  -2.36896321  0.56503692   0   0
#>   0.56503692  -3.95890385  0.57618390   0   1
#>   0.30481963   0.19998186  0.32649266   0   0
#>   0.32649266   0.22763331  0.39255439   0   0
#>   0.39255439   0.10823663  0.39608667   0   0
#>   0.39608667   0.07810493  0.34555463   0   0
#>   0.34555463   0.08889577  0.31865440   0   0
#>   0.31865440   0.12714460  0.30481963   0   0
#>   0.83134928   0.85417958  0.81261156   0   0
#>   0.81261156   0.89986096  0.70637089   0   0
#>   0.70637089   0.98446481  0.63355157   0   0
#>   0.63355157   1.02114308  0.62305081   0   0
#>   0.62305081   2.51851103  0.54570067   0   0
#>   0.54570067   2.85973429  0.53486504   0   0
#>   0.53486504   3.83961266  0.54425103   0   0
#>   0.54425103   5.67644099  0.57777806   0   1
#>   0.24416975   0.93733544  0.18054364   0   0
#>   0.18054364   0.94083691  0.11708343   0   0
#>   0.11708343   4.67446273 -1.82595091   0   1
#>   0.29194363   0.58559732  0.23623288   0   0
#>   0.23623288   0.59492769  0.18006560   0   0
#>   0.18006560   0.66291959  0.15049642   0   0
#>   0.15049642   0.67899875  0.15096025   0   0
#>   0.15096025   0.72261808  0.27056632   0   0
#>   0.27056632   0.59879140  0.29194363   0   0
#>   0.47280578   0.48446558  0.41603367   0   0
#>   0.41603367   0.48292421  0.40998871   0   0
#>   0.40998871   0.55632237  0.42654279   0   0
#>   0.42654279   0.52971316  0.47280578   0   0
#>   0.79670720   0.72634185  0.69326900   0   0
#>   0.69326900   0.71570328  0.67799047   0   0
#>   0.67799047   0.80987614  0.64315005   0   0
#>   0.64315005   0.84521379  0.66088367   0   0
#>   0.58295382  -0.42414189  0.58506899   0   0
#>   0.58506899  -0.86567584  0.56791416   0   0
#>   0.40327313   0.02173345  0.48222860   0   0
#>   0.48222860  -0.32003516  0.58295382   0   0
#>   0.40418547   0.60220014  0.44296825   0   0
#>   0.44296825   0.57670247  0.48585487   0   0
#>   0.48585487   0.54619063  0.50388651   0   0
#>   1.10639799   0.25044238  0.96845155   0   0
#>   0.91240690   0.31267467  1.10639799   0   0
#>   0.10657248   0.78754038  0.06902652   0   0
#>   0.06902652   0.67207466 -0.66318998   0   0
#>  -0.66318998   0.71602692 -2.81873459   0   1
#>   0.82913476   0.11496183  0.80365321   0   0
#>   0.80365321   0.12179519  0.78641714   0   0
#>   0.78641714   0.12485421  0.78279388   0   0
#>   0.78279388   0.23481125  0.71463583   0   0
#>   0.34139067   0.90609772  0.39840990   0   0
#>   0.39840990   0.82376397  0.46287508   0   0
#>   0.46287508   0.75214847  0.43310832   0   0
#>   0.43310832   0.73589178  0.42165988   0   0
#>   0.42165988   0.73461272  0.41608752   0   0
#>   0.41608752   0.78001121  0.29670506   0   0
#>   0.29670506   0.78198671  0.29598099   0   0
#>   0.29598099   0.88520959  0.31804513   0   0
#>   0.31804513   0.91234755  0.34139067   0   0
#>   0.48298479   0.13404977  0.48200232   0   0
#>   0.48200232   0.20234621  0.54144746   0   0
#>   0.38814141   0.59930002  0.36564351   0   0
#>   0.36564351   0.59588938  0.29616650   0   0
#>   0.10003659   0.57014723  0.01689095   0   0
#>   0.01689095   0.55991027 -0.03872836   0   0
#>   0.54085166   0.28091922  0.57042085   0   0
#>   0.57042085   0.26887028  0.65380783   0   0
#>   0.65380783   0.23487205  0.71425720   0   0
#>   0.56250239   0.78341133  0.56539500   0   0
#>   0.56539500   0.68574336  0.58337768   0   0
#>   0.58337768   0.68184538  0.52914203   0   0
#>   0.52914203   0.68946223  0.50008127   0   0
#>   0.27423813   0.29711497  0.26105863   0   0
#>   0.26105863   0.32154930  0.17336259   0   0
#>   0.17336259   0.43683323  0.16180316   0   0
#>   0.16180316   0.44289738  0.16343833   0   0
#>   0.16343833   0.44412257  0.16476160   0   0
#>   0.16476160   0.44428133  0.18264351   0   0
#>   0.18264351   0.40238717  0.26295798   0   0
#>   0.26295798   0.35536307  0.28674817   0   0
#>   0.28674817   0.31134126  0.27423813   0   0
#>   0.88648615   0.46445589  0.91216592   0   0
#>   0.72425344   0.06147748  0.72664110   0   0
#>   0.72664110  -0.17957368  0.61986661   0   0
#>   0.61986661   0.06480189  0.69003990   0   0
#>   0.69003990   0.07432381  0.72425344   0   0
#>   0.02881679   0.26325166 -0.19254830   0   0
#>  -0.19254830   0.37236973  0.02837460   0   0
#>   0.02837460   0.34883245  0.07501305   0   0
#>   0.07501305   0.23547978  0.02881679   0   0
#>   0.59830061   0.42369978  0.62206272   0   0
#>   0.62206272   0.36885079  0.72057705   0   0
#>   0.91305378   0.62440036  1.76359485   0   0
#>   1.76359485   0.63058119  1.15056568   0   0
#>   1.15056568   0.73006525  0.92681419   0   0
#>   0.92681419   0.80930455  0.91305378   0   0
#>   0.32249572   0.46965022  0.39447738   0   0
#>   0.39447738   0.37745974  0.40537712   0   0
#>   0.91059028   0.07124890  0.97604909   0   0
#>   0.97604909   0.08293963  1.13481655   0   0
#>   1.13481655  -1.10950237  7.08819748   0   1
#>   0.24509787   0.23003945  0.22960834   0   0
#>   0.22960834   0.25343640  0.24894424   0   0
#>   0.95388173   0.10408263  0.93652239   0   0
#>   0.93652239   0.07842424  0.86678761   0   0
#>   0.30001874   0.47893531  0.20914295   0   0
#>   3.24798081   0.49362159  2.97751682   0   0
#>   0.58229400   0.89436219  0.56867409   0   0
#>   0.83279246   0.65351483  0.88738424   0   0
#>   0.11287362   0.22005268  0.10131448   0   0
#>   0.80084467  -0.08068039  0.79448648   0   0
#>   0.79448648  -0.94179666  0.57103780   0   0
#>   0.47039036   0.87269136  0.51787557   0   0
#>   0.51787557   0.83899993  0.50261822   0   0
#>   0.50152114   0.44979263  0.50075535   0   0
#>   0.50075535   0.52935015  0.54746789   0   0
#>   0.54746789   0.54173397  0.64854509   0   0
#>   0.12206893  -1.39801211 -6.89184283   0   1
#>   0.20293721 -16.17910160  0.26040816   0   1
#>   0.83134928   6.99975053  1.52889593   0   1
#>   0.91059028  -9.01720877  3.39515253   0   1
#>   0.87044434   4.24864965  6.69429154   0   1
#> 
#> $x
#>                [,1]        [,2]
#>   [1,] 0.2394844432 0.868392250
#>   [2,] 0.1022727633 0.699891507
#>   [3,] 0.0739201886 0.541939437
#>   [4,] 0.0139627554 0.062379183
#>   [5,] 0.4201805218 0.772351362
#>   [6,] 0.0002177514 0.233949568
#>   [7,] 0.1698106904 0.352463049
#>   [8,] 0.9900662483 0.764447526
#>   [9,] 1.0000000000 0.220212154
#>  [10,] 0.6468548765 0.216160065
#>  [11,] 0.5440356441 0.431865236
#>  [12,] 0.7956553358 0.680854684
#>  [13,] 0.0405854275 0.429518677
#>  [14,] 0.5578998374 0.439839542
#>  [15,] 0.3097111134 0.945928442
#>  [16,] 0.8903191318 0.009455548
#>  [17,] 0.2057182720 0.799959867
#>  [18,] 0.8290425672 0.407903004
#>  [19,] 0.0937539928 0.526214514
#>  [20,] 0.9792669739 0.424368998
#>  [21,] 0.6694818991 0.347225637
#>  [22,] 0.6714172162 0.043974941
#>  [23,] 0.1743181222 0.647460371
#>  [24,] 0.7543268155 0.513998775
#>  [25,] 0.3547198093 0.199025547
#>  [26,] 0.4076992357 0.847323419
#>  [27,] 0.0477042865 0.718941167
#>  [28,] 0.3168822633 0.054187774
#>  [29,] 0.3613517739 0.674485659
#>  [30,] 0.6506667293 0.084015930
#>  [31,] 0.3845642340 0.088345202
#>  [32,] 0.7419853025 0.977904195
#>  [33,] 0.5475097436 0.416461632
#>  [34,] 0.4256538952 0.339235045
#>  [35,] 0.0328407031 0.965632656
#>  [36,] 0.1889168710 0.288252286
#>  [37,] 0.1227444685 0.921630979
#>  [38,] 0.5218437942 0.245766664
#>  [39,] 0.5226810541 0.975693076
#>  [40,] 0.0000000000 0.172395956
#>  [41,] 0.8558616887 0.608524306
#>  [42,] 0.5847189891 0.907511407
#>  [43,] 0.3489642778 0.141624779
#>  [44,] 0.6039356686 0.117470887
#>  [45,] 0.5276516604 0.354188107
#>  [46,] 0.2798874858 0.178175370
#>  [47,] 0.0656429512 0.656470509
#>  [48,] 0.7005573659 0.758794839
#>  [49,] 0.4080477800 0.977311861
#>  [50,] 0.8706870592 0.213077199
#>  [51,] 0.0758834074 0.468032982
#>  [52,] 0.0465200416 0.721614686
#>  [53,] 0.2396567806 0.323227575
#>  [54,] 0.4851568267 0.078256923
#>  [55,] 0.0271155754 0.776207895
#>  [56,] 0.0256230671 0.820053103
#>  [57,] 0.8316069445 0.609214194
#>  [58,] 0.9948071196 0.269509899
#>  [59,] 0.0943717650 0.705953209
#>  [60,] 0.6841477264 0.952188535
#>  [61,] 0.8783971552 0.470937787
#>  [62,] 0.0070477820 0.287171795
#>  [63,] 0.1077171012 0.960119183
#>  [64,] 0.4475415194 0.544171708
#>  [65,] 0.1722818956 0.435993917
#>  [66,] 0.0802195305 0.937277715
#>  [67,] 0.1870395163 0.968840270
#>  [68,] 0.7745397451 0.623779716
#>  [69,] 0.9199459485 0.519523319
#>  [70,] 0.3234034372 0.731309665
#>  [71,] 0.4119797922 0.702673390
#>  [72,] 0.0046270986 0.862872955
#>  [73,] 0.1454525876 0.702735307
#>  [74,] 0.9717863502 0.925294758
#>  [75,] 0.4866092529 0.477634283
#>  [76,] 0.4397495373 0.458456613
#>  [77,] 0.0513907837 0.738775688
#>  [78,] 0.8436238327 1.000000000
#>  [79,] 0.1294600918 0.240906188
#>  [80,] 0.0818577427 0.512047739
#>  [81,] 0.4820943447 0.265467497
#>  [82,] 0.9756621670 0.485608867
#>  [83,] 0.1943335944 0.997979134
#>  [84,] 0.6906508865 0.753970156
#>  [85,] 0.0661583775 0.516289714
#>  [86,] 0.3086837246 0.361022733
#>  [87,] 0.1503327543 0.033292910
#>  [88,] 0.6079433495 0.524519538
#>  [89,] 0.3114272231 0.067572913
#>  [90,] 0.6261101441 0.480393071
#>  [91,] 0.7885212911 0.505392679
#>  [92,] 0.9656675457 0.249083632
#>  [93,] 0.5341196009 0.197432756
#>  [94,] 0.6460660465 0.449902757
#>  [95,] 0.7240626520 0.874696574
#>  [96,] 0.4019415805 0.518918745
#>  [97,] 0.4265920520 0.000000000
#>  [98,] 0.7455692613 0.180159755
#>  [99,] 0.8411777127 0.553126120
#> [100,] 0.1644478942 0.200834633
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.2394844,0.8683923) [6]: 15 17 26 37 67 70 
#> 2: (0.1022728,0.6998915) [4]: 23 47 59 73 
#> 3: (0.07392019,0.5419394) [5]: 19 23 47 80 85 
#> 4: (0.01396276,0.06237918) [3]: 40 87 100 
#> 5: (0.4201805,0.7723514) [5]: 26 42 70 71 84 
#> 6: (0.0002177514,0.2339496) [4]: 40 62 72 79 
#> 7: (0.1698107,0.352463) [6]: 13 36 53 62 65 79 
#> 8: (0.9900662,0.7644475) [9]: 9 12 20 41 58 69 74 82 95 
#> 9: (1,0.2202122) [5]: 8 16 50 58 92 
#> 10: (0.6468549,0.2161601) [6]: 21 30 38 44 93 98 
#> 11: (0.5440356,0.4318652) [4]: 14 33 75 76 
#> 12: (0.7956553,0.6808547) [7]: 8 41 48 57 68 84 95 
#> 13: (0.04058543,0.4295187) [7]: 7 51 52 55 62 65 85 
#> 14: (0.5578998,0.4398395) [6]: 11 33 75 88 90 94 
#> 15: (0.3097111,0.9459284) [5]: 1 26 49 67 83 
#> 16: (0.8903191,0.009455548) [5]: 9 22 50 97 98 
#> 17: (0.2057183,0.7999599) [7]: 1 37 56 59 70 73 77 
#> 18: (0.8290426,0.407903) [9]: 20 21 24 50 61 91 92 94 98 
#> 19: (0.09375399,0.5262145) [5]: 3 23 51 65 80 
#> 20: (0.979267,0.424369) [6]: 8 18 58 61 82 92 
#> 21: (0.6694819,0.3472256) [7]: 10 18 33 38 45 94 98 
#> 22: (0.6714172,0.04397494) [5]: 16 30 54 97 98 
#> 23: (0.1743181,0.6474604) [9]: 2 3 19 29 47 65 70 73 96 
#> 24: (0.7543268,0.5139988) [7]: 18 68 88 90 91 94 99 
#> 25: (0.3547198,0.1990255) [9]: 31 34 43 46 53 54 81 86 93 
#> 26: (0.4076992,0.8473234) [7]: 1 5 15 39 42 49 70 
#> 27: (0.04770429,0.7189412) [4]: 47 52 59 77 
#> 28: (0.3168823,0.05418777) [4]: 31 87 89 97 
#> 29: (0.3613518,0.6744857) [5]: 23 64 70 71 96 
#> 30: (0.6506667,0.08401593) [5]: 10 22 44 54 98 
#> 31: (0.3845642,0.0883452) [6]: 25 28 43 54 89 97 
#> 32: (0.7419853,0.9779042) [4]: 39 60 78 95 
#> 33: (0.5475097,0.4164616) [6]: 11 14 21 45 76 94 
#> 34: (0.4256539,0.339235) [5]: 25 45 76 81 86 
#> 35: (0.0328407,0.9656327) [4]: 63 66 72 83 
#> 36: (0.1889169,0.2882523) [5]: 7 46 53 79 100 
#> 37: (0.1227445,0.921631) [6]: 1 17 56 63 66 67 
#> 38: (0.5218438,0.2457667) [5]: 10 21 45 81 93 
#> 39: (0.5226811,0.9756931) [7]: 26 32 42 49 60 78 83 
#> 40: (0,0.172396) [4]: 4 6 79 100 
#> 41: (0.8558617,0.6085243) [5]: 8 12 57 69 99 
#> 42: (0.584719,0.9075114) [6]: 5 26 39 60 84 95 
#> 43: (0.3489643,0.1416248) [4]: 25 31 46 89 
#> 44: (0.6039357,0.1174709) [4]: 10 30 54 93 
#> 45: (0.5276517,0.3541881) [6]: 21 33 34 38 76 81 
#> 46: (0.2798875,0.1781754) [7]: 25 36 43 53 87 89 100 
#> 47: (0.06564295,0.6564705) [7]: 2 3 23 27 52 59 85 
#> 48: (0.7005574,0.7587948) [3]: 12 84 95 
#> 49: (0.4080478,0.9773119) [4]: 15 26 39 83 
#> 50: (0.8706871,0.2130772) [5]: 9 16 18 92 98 
#> 51: (0.07588341,0.468033) [5]: 13 19 65 80 85 
#> 52: (0.04652004,0.7216147) [6]: 13 27 47 55 77 85 
#> 53: (0.2396568,0.3232276) [6]: 7 25 36 46 65 86 
#> 54: (0.4851568,0.07825692) [7]: 22 25 30 31 44 93 97 
#> 55: (0.02711558,0.7762079) [6]: 13 52 56 62 72 77 
#> 56: (0.02562307,0.8200531) [6]: 17 37 55 66 72 77 
#> 57: (0.8316069,0.6092142) [4]: 12 41 68 99 
#> 58: (0.9948071,0.2695099) [4]: 8 9 20 92 
#> 59: (0.09437176,0.7059532) [6]: 2 17 27 47 73 77 
#> 60: (0.6841477,0.9521885) [4]: 32 39 42 95 
#> 61: (0.8783972,0.4709378) [6]: 18 20 69 82 91 99 
#> 62: (0.007047782,0.2871718) [6]: 6 7 13 55 72 79 
#> 63: (0.1077171,0.9601192) [5]: 35 37 66 67 83 
#> 64: (0.4475415,0.5441717) [6]: 29 71 75 76 88 96 
#> 65: (0.1722819,0.4359939) [8]: 7 13 19 23 51 53 86 96 
#> 66: (0.08021953,0.9372777) [5]: 35 37 56 63 72 
#> 67: (0.1870395,0.9688403) [5]: 1 15 37 63 83 
#> 68: (0.7745397,0.6237797) [6]: 12 24 57 84 88 99 
#> 69: (0.9199459,0.5195233) [5]: 8 41 61 82 99 
#> 70: (0.3234034,0.7313097) [8]: 1 5 17 23 26 29 71 73 
#> 71: (0.4119798,0.7026734) [6]: 5 29 64 70 84 88 
#> 72: (0.004627099,0.862873) [6]: 6 35 55 56 62 66 
#> 73: (0.1454526,0.7027353) [5]: 2 17 23 59 70 
#> 74: (0.9717864,0.9252948) [3]: 8 78 95 
#> 75: (0.4866093,0.4776343) [5]: 11 14 64 76 88 
#> 76: (0.4397495,0.4584566) [8]: 11 33 34 45 64 75 86 96 
#> 77: (0.05139078,0.7387757) [6]: 17 27 52 55 56 59 
#> 78: (0.8436238,1) [5]: 32 39 74 83 95 
#> 79: (0.1294601,0.2409062) [6]: 6 7 36 40 62 100 
#> 80: (0.08185774,0.5120477) [4]: 3 19 51 85 
#> 81: (0.4820943,0.2654675) [5]: 25 34 38 45 93 
#> 82: (0.9756622,0.4856089) [4]: 8 20 61 69 
#> 83: (0.1943336,0.9979791) [7]: 15 35 39 49 63 67 78 
#> 84: (0.6906509,0.7539702) [8]: 5 12 42 48 68 71 88 95 
#> 85: (0.06615838,0.5162897) [6]: 3 13 47 51 52 80 
#> 86: (0.3086837,0.3610227) [6]: 25 34 53 65 76 96 
#> 87: (0.1503328,0.03329291) [6]: 4 28 46 89 97 100 
#> 88: (0.6079433,0.5245195) [8]: 14 24 64 68 71 75 84 90 
#> 89: (0.3114272,0.06757291) [5]: 28 31 43 46 87 
#> 90: (0.6261101,0.4803931) [4]: 14 24 88 94 
#> 91: (0.7885213,0.5053927) [4]: 18 24 61 99 
#> 92: (0.9656675,0.2490836) [5]: 9 18 20 50 58 
#> 93: (0.5341196,0.1974328) [6]: 10 25 38 44 54 81 
#> 94: (0.646066,0.4499028) [6]: 14 18 21 24 33 90 
#> 95: (0.7240627,0.8746966) [9]: 8 12 32 42 48 60 74 78 84 
#> 96: (0.4019416,0.5189187) [6]: 23 29 64 65 76 86 
#> 97: (0.4265921,0) [6]: 16 22 28 31 54 87 
#> 98: (0.7455693,0.1801598) [7]: 10 16 18 21 22 30 50 
#> 99: (0.8411777,0.5531261) [7]: 24 41 57 61 68 69 91 
#> 100: (0.1644479,0.2008346) [6]: 4 36 40 46 79 87 
#> number of nodes: 100 
#> number of arcs: 284 
#> number of boundary nodes: 13 
#> boundary nodes:  4 6 8 9 16 35 40 72 74 78 83 87 97 
#> number of triangles: 185 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.079765722 0.147270199 0.136244810 0.143404926 0.139013031 0.124792919 
#>                                                                         
#> 0.057217850 0.109048042 0.062358141 0.051182080 0.140181743 0.016262362 
#>                                                                         
#> 0.018499143 0.145571758 0.168135262 0.166888790 0.093090316 0.283353026 
#>                                                                         
#> 0.186538153 0.186972156 0.146470363 0.130863593 0.575706249 0.076005010 
#>                                                                         
#> 0.061626368 0.131360660 0.131855559 0.146244345 0.252838302 0.206640245 
#>                                                                         
#> 0.094108918 0.109672093 0.065226429 0.061345874 0.155636874 0.049570492 
#>                                                                         
#> 0.035585772 0.101576355 0.204565870 0.179139424 0.049868448 0.106151702 
#>                                                                         
#> 0.057471960 0.135093026 0.183576402 0.080689466 0.050632163 0.115653969 
#>                                                                         
#> 0.025582821 0.118261197 0.011018888 0.154877020 0.058896687 0.024264554 
#>                                                                         
#> 0.206261990 0.057939208 0.489447672 0.101583317 0.049226906 0.104075231 
#>                                                                         
#> 0.036440324 0.047719782 0.130077631 0.214711176 0.030037926 0.129988909 
#>                                                                         
#> 0.129375506 0.155063952 0.248743154 0.140536940 0.085264256 0.054080310 
#>                                                                         
#> 0.180230405 0.175812127 0.111100731 0.096253886 0.035260843 0.125812136 
#>                                                                         
#> 0.105313310 0.044418369 0.211476256 0.098365254 0.065363105 0.108576888 
#>                                                                         
#> 0.165610066 0.097744456 0.244172271 0.160774990 0.068330403 0.071071655 
#>                                                                         
#> 0.097121313 0.193925094 0.078696842 0.150527911 0.078150809 0.064078625 
#>                                                                         
#> 0.101097196 0.128841890 0.085646677 0.092882590 0.118982033 0.092181589 
#>                                                                         
#> 0.114644703 0.017838838 0.067892927 0.164701386 0.278258185 0.097832573 
#>                                                                         
#> 0.076029668 0.052125489 0.162442050 0.125084564 0.083022342 0.173193027 
#>                                                                         
#> 0.321861917 0.163168470 0.087167685 0.136377068 0.163314445 0.099733723 
#>                                                                         
#> 0.106067356 0.035747015 0.094529323 0.053196577 0.117642431 0.129315055 
#>                                                                         
#> 0.079800394 0.649293383 0.109005135 0.056898784 0.085636343 0.125264576 
#>                                                                         
#> 0.115011557 0.044614525 0.155647478 0.047690361 0.264840538 0.071309964 
#>                                                                         
#> 0.183379874 0.243901871 0.128306601 0.076309537 0.160729688 0.169529089 
#>                                                                         
#> 0.104611884 0.113314937 0.056807801 0.089104196 0.043273369 0.009958414 
#>                                                                         
#> 0.026798396 0.030927631 0.025311124 0.145651682 0.114829783 0.204488847 
#>                                                                         
#> 0.110899274 0.070158904 0.271094240 0.212934603 0.076003894 0.105120050 
#>                                                                         
#> 0.129429429 0.053658688 0.628938844 0.066993046 0.075717827 0.083567415 
#>                                                                         
#> 0.150455104 0.175370215 0.118630115 0.287945769 0.211620807 0.205725337 
#>                                                                         
#> 0.254764054 0.279210450 0.340249951 0.494960332 0.544326023 0.044858440 
#>                                                                         
#> 0.129509630 0.237588383 0.128469146 0.114280158 0.107617899 0.132199102 
#>                                                                         
#> 0.105074031 0.133004384 0.073434298 0.107622922 0.015790515 0.015993918 
#>                                                                         
#> 0.122956462 0.127952361 0.060855731 0.080155337 0.292156291 0.346950791 
#>                                                                         
#> 0.052242694 0.090460978 0.088738654 0.079355117 0.098361857 0.126575062 
#>                                                                         
#> 0.103223151 0.223813429 0.221606943 0.463823471 0.181212640 0.166013485 
#>                                                                         
#> 0.145723355 0.114387795 0.151124126 0.080057849 0.105575615 0.129764233 
#>                                                                         
#> 0.187734883 0.170708376 0.242558818 0.199226901 0.209499350 0.060864180 
#>                                                                         
#> 0.119609312 0.140252736 0.142001030 0.045098375 0.189388983 0.261410383 
#>                                                                         
#> 0.188976065 0.171047173 0.095257690 0.111626241 0.146761049 0.132547569 
#>                                                                         
#> 0.169309316 0.077682730 0.057688598 0.114633452 0.177760727 0.179406862 
#>                                                                         
#> 0.143662121 0.157131627 0.168411429 0.172335682 0.020174204 0.002924062 
#>                                                                         
#> 0.064995222 0.048441104 0.167855099 0.122362382 0.075812790 0.014454045 
#>                                                                         
#> 0.156238273 0.057946047 0.104012573 0.219315395 0.063296764 0.104752252 
#>                                                                         
#> 0.103088012 0.120051944 0.055215541 0.075079114 0.164700482 0.090777623 
#>                                                                         
#> 0.142802592 0.041317836 0.045312147 0.044363742 0.329102906 0.057311239 
#>                                                                         
#> 0.143155404 0.194358832 0.043870603 0.089535268 0.063928524 0.090223107 
#>                                                                         
#> 0.086068534 0.077159036 0.161601222 0.139437398 0.061553997 0.161882633 
#>                         
#> 0.106562486 0.148345893 
#> 
#> $alpha
#> [1] 0.2960105
#> 
#> attr(,"class")
#> [1] "scree"