Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1         y1           x2         y2         mx1
#>    78   70 0.2817090879 0.67905268 0.2674387529 0.59953300  0.31941447
#>     1   70 0.3412471555 0.57440827 0.2674387529 0.59953300  0.28319567
#>    12   70 0.2283720613 0.57795693 0.2674387529 0.59953300  0.21727813
#>    78    1 0.2817090879 0.67905268 0.3412471555 0.57440827  0.35310662
#>    13   78 0.1502484818 0.63992554 0.2817090879 0.67905268  0.20294058
#>    78   54 0.2817090879 0.67905268 0.4294710582 0.64113181  0.36066396
#>    52   87 0.4370319794 0.64584735 0.4111831543 0.76588825  0.39787139
#>     1    3 0.3412471555 0.57440827 0.3769197176 0.44418620  0.39672960
#>    87   79 0.4111831543 0.76588825 0.3162584933 0.74515231  0.37648614
#>    56   41 0.3272437655 0.84932830 0.2474956415 0.91017972  0.23267298
#>    54    1 0.4294710582 0.64113181 0.3412471555 0.57440827  0.41929741
#>     1   12 0.3412471555 0.57440827 0.2283720613 0.57795693  0.28217403
#>    54   32 0.4294710582 0.64113181 0.4734235371 0.50549592  0.46942632
#>    52   54 0.4370319794 0.64584735 0.4294710582 0.64113181  0.47105774
#>    56   55 0.3272437655 0.84932830 0.3355347859 0.79999857  0.37952937
#>    48   87 0.4653866597 0.77354713 0.4111831543 0.76588825  0.44662179
#>    24   40 0.1256600354 0.74130097 0.1215670361 0.69023665  0.20336527
#>    32    3 0.4734235371 0.50549592 0.3769197176 0.44418620  0.39851828
#>    41   79 0.2474956415 0.91017972 0.3162584933 0.74515231  0.23029838
#>    55   87 0.3355347859 0.79999857 0.4111831543 0.76588825  0.39108944
#>    79   24 0.3162584933 0.74515231 0.1256600354 0.74130097  0.22076202
#>    52   76 0.4370319794 0.64584735 0.5420669327 0.59189760  0.50580932
#>    53   76 0.5161512630 0.49853860 0.5420669327 0.59189760  0.50300180
#>    48   52 0.4653866597 0.77354713 0.4370319794 0.64584735  0.46333936
#>    29   53 0.4471462943 0.45871028 0.5161512630 0.49853860  0.48892552
#>     3   18 0.3769197176 0.44418620 0.2211755277 0.41171380  0.31198948
#>    13   68 0.1502484818 0.63992554 0.0463988403 0.59373594  0.13326094
#>    40   44 0.1215670361 0.69023665 0.0922670192 0.72170746  0.05905043
#>    13   40 0.1502484818 0.63992554 0.1215670361 0.69023665  0.08880613
#>    16   53 0.5353357860 0.49610453 0.5161512630 0.49853860  0.51587044
#>    48   26 0.4653866597 0.77354713 0.5513284556 0.74173575  0.54898224
#>    48   99 0.4653866597 0.77354713 0.5293945274 0.71530522  0.50763213
#>    99   63 0.5293945274 0.71530522 0.6033538354 0.61577613  0.56915768
#>    68   18 0.0463988403 0.59373594 0.2211755277 0.41171380  0.12985014
#>    76   63 0.5420669327 0.59189760 0.6033538354 0.61577613  0.55264447
#>    76   16 0.5420669327 0.59189760 0.5353357860 0.49610453  0.59763706
#>    41   45 0.2474956415 0.91017972 0.1692505674 0.91081736  0.20755902
#>    80   41 0.4165815882 0.97346585 0.2474956415 0.91017972  0.33292465
#>    26   10 0.5513284556 0.74173575 0.6306865955 0.67598905  0.63552503
#>    26   99 0.5513284556 0.74173575 0.5293945274 0.71530522  0.58002437
#>    69   44 0.0469287132 0.99306009 0.0922670192 0.72170746  0.03818292
#>    51   93 0.4549301015 0.34317164 0.5388787784 0.34389511  0.49638337
#>    51   29 0.4549301015 0.34317164 0.4471462943 0.45871028  0.42287125
#>    51    3 0.4549301015 0.34317164 0.3769197176 0.44418620  0.37556471
#>    47   80 0.4078680355 0.87971615 0.4165815882 0.97346585  0.34169083
#>    47   48 0.4078680355 0.87971615 0.4653866597 0.77354713  0.53398183
#>    93   57 0.5388787784 0.34389511 0.6423125597 0.46359956  0.57087688
#>    57   63 0.6423125597 0.46359956 0.6033538354 0.61577613  0.60567024
#>     2   93 0.5659857850 0.29180963 0.5388787784 0.34389511  0.53275401
#>    72   10 0.7161990718 0.50276073 0.6306865955 0.67598905  0.70604762
#>    18   86 0.2211755277 0.41171380 0.1207797805 0.32297509  0.19767807
#>    22   86 0.0000000000 0.36564842 0.1207797805 0.32297509  0.09778489
#>    95   62 0.3682819109 0.28114786 0.3351649278 0.25392267  0.41692195
#>    62   18 0.3351649278 0.25392267 0.2211755277 0.41171380  0.28962647
#>    68   22 0.0463988403 0.59373594 0.0000000000 0.36564842  0.09466010
#>    57   72 0.6423125597 0.46359956 0.7161990718 0.50276073  0.64614284
#>    27   57 0.6869975704 0.37719243 0.6423125597 0.46359956  0.60997599
#>    51   95 0.4549301015 0.34317164 0.3682819109 0.28114786  0.44860213
#>    97   94 0.6422775611 0.96109647 0.6220334333 0.96966175  0.58585780
#>    39   59 0.7537823029 0.73139852 0.7142486443 0.79902179  0.67901701
#>    62   34 0.3351649278 0.25392267 0.2075239239 0.20126346  0.29614455
#>    62   81 0.3351649278 0.25392267 0.2274968009 0.26485422  0.27676795
#>    86   82 0.1207797805 0.32297509 0.1026731056 0.23936609  0.16134520
#>    39   10 0.7537823029 0.73139852 0.6306865955 0.67598905  0.72351859
#>    28    2 0.6809128051 0.35860639 0.5659857850 0.29180963  0.63184876
#>    30   71 0.5450761138 0.27293833 0.5729773120 0.26301990  0.56171003
#>    30   51 0.5450761138 0.27293833 0.4549301015 0.34317164  0.47194095
#>    82   34 0.1026731056 0.23936609 0.2075239239 0.20126346  0.16564903
#>    68   44 0.0463988403 0.59373594 0.0922670192 0.72170746 -0.31052427
#>    45   69 0.1692505674 0.91081736 0.0469287132 0.99306009  0.04173334
#>    41   15 0.2474956415 0.91017972 0.0551180008 0.99962689  0.20975949
#>    15   69 0.0551180008 0.99962689 0.0469287132 0.99306009  0.09830279
#>    94   88 0.6220334333 0.96966175 0.8273331347 1.00000000  0.54239345
#>    94   80 0.6220334333 0.96966175 0.4165815882 0.97346585  0.51829227
#>    59   97 0.7142486443 0.79902179 0.6422775611 0.96109647  0.60557266
#>   100   71 0.5182220408 0.21540067 0.5729773120 0.26301990  0.54787132
#>    21   96 0.7954210508 0.92711351 0.7029433401 0.92029182  0.75359713
#>    59   96 0.7142486443 0.79902179 0.7029433401 0.92029182  0.61204979
#>    96   89 0.7029433401 0.92029182 0.6677166753 0.95503520  0.63892554
#>    35   50 0.0478217700 0.21653189 0.1609054711 0.08237208  0.10700005
#>    73   90 0.2031979687 0.14861514 0.5095829883 0.10875321  0.35228480
#>    73   34 0.2031979687 0.14861514 0.2075239239 0.20126346  0.14054451
#>    72    6 0.7161990718 0.50276073 0.7783675188 0.43791470  0.80731937
#>    72   27 0.7161990718 0.50276073 0.6869975704 0.37719243  0.71287832
#>     6   27 0.7783675188 0.43791470 0.6869975704 0.37719243  0.75026579
#>     2   71 0.5659857850 0.29180963 0.5729773120 0.26301990  0.62718902
#>    71   60 0.5729773120 0.26301990 0.6008289234 0.20247516  0.62750414
#>     5   91 0.8162889544 0.67612565 0.8135425629 0.73571539  0.78575587
#>    39   91 0.7537823029 0.73139852 0.8135425629 0.73571539  0.77945690
#>   100   60 0.5182220408 0.21540067 0.6008289234 0.20247516  0.56141647
#>    25   60 0.5732349509 0.17134203 0.6008289234 0.20247516  0.55984844
#>    31   92 0.6818415372 0.26387744 0.7263512358 0.26693038  0.70093919
#>    50   73 0.1609054711 0.08237208 0.2031979687 0.14861514  0.11862683
#>     8   74 0.0001700435 0.18619815 0.0266994967 0.19763402 -0.02133424
#>    25   11 0.5732349509 0.17134203 0.5966151683 0.16936733  0.58637257
#>    90  100 0.5095829883 0.10875321 0.5182220408 0.21540067  0.41313668
#>    14   31 0.7047697002 0.18843387 0.6818415372 0.26387744  0.65539008
#>    25   90 0.5732349509 0.17134203 0.5095829883 0.10875321  0.54434286
#>    27   28 0.6869975704 0.37719243 0.6809128051 0.35860639  0.74205902
#>    22    8 0.0000000000 0.36564842 0.0001700435 0.18619815 -0.02345158
#>     5   72 0.8162889544 0.67612565 0.7161990718 0.50276073  0.76976777
#>     6   65 0.7783675188 0.43791470 0.8039568121 0.34740036  0.84043587
#>    25  100 0.5732349509 0.17134203 0.5182220408 0.21540067  0.52025346
#>    88   89 0.8273331347 1.00000000 0.6677166753 0.95503520  0.74244003
#>    21   88 0.7954210508 0.92711351 0.8273331347 1.00000000  0.74406846
#>    91   21 0.8135425629 0.73571539 0.7954210508 0.92711351  0.80474494
#>    92   65 0.7263512358 0.26693038 0.8039568121 0.34740036  0.74043160
#>    35   74 0.0478217700 0.21653189 0.0266994967 0.19763402  0.09550389
#>    25   77 0.5732349509 0.17134203 0.5537893075 0.09324029  0.56863913
#>    90   19 0.5095829883 0.10875321 0.1547900962 0.05954129  0.36355898
#>    84   75 0.8905573427 0.20356443 0.8862984169 0.32953133  0.82908522
#>    11   14 0.5966151683 0.16936733 0.7047697002 0.18843387  0.65061918
#>    50   74 0.1609054711 0.08237208 0.0266994967 0.19763402  0.05757085
#>    11   85 0.5966151683 0.16936733 0.5778672680 0.09148920  0.62688754
#>     9   85 0.5938670025 0.06413792 0.5778672680 0.09148920  0.56383264
#>    23    6 0.8692314801 0.59931976 0.7783675188 0.43791470  0.84160276
#>     7   65 0.9032046093 0.43648666 0.8039568121 0.34740036  0.85611915
#>    14   64 0.7047697002 0.18843387 0.6826177784 0.13357991  0.71737713
#>    88   15 0.8273331347 1.00000000 0.0551180008 0.99962689  0.43986995
#>     5   23 0.8162889544 0.67612565 0.8692314801 0.59931976  0.92124889
#>    23   58 0.8692314801 0.59931976 0.9470835631 0.50049843  1.00418012
#>    20    9 0.6023597247 0.02411119 0.5938670025 0.06413792  0.53973465
#>     7   58 0.9032046093 0.43648666 0.9470835631 0.50049843  0.86079916
#>    92   61 0.7263512358 0.26693038 0.8253213439 0.14532774  0.80700779
#>    14   43 0.7047697002 0.18843387 0.7232782755 0.11281024  0.76356762
#>    42   43 0.6436831113 0.05509317 0.7232782755 0.11281024  0.68314131
#>    64   85 0.6826177784 0.13357991 0.5778672680 0.09148920  0.63118274
#>    20   42 0.6023597247 0.02411119 0.6436831113 0.05509317  0.61667763
#>     4   46 0.9612747158 0.80272784 0.9379403792 0.91116179  0.85750121
#>    20   37 0.6023597247 0.02411119 0.6166264897 0.00000000  0.63317038
#>    90   20 0.5095829883 0.10875321 0.6023597247 0.02411119  0.49679454
#>    91    4 0.8135425629 0.73571539 0.9612747158 0.80272784  0.85694435
#>    33   83 1.0000000000 0.76296818 0.9659560170 0.85710218  1.02365448
#>    98   43 0.7346105442 0.03412311 0.7232782755 0.11281024  0.69466371
#>    92   84 0.7263512358 0.26693038 0.8905573427 0.20356443  0.81693594
#>    14   61 0.7047697002 0.18843387 0.8253213439 0.14532774  0.78041103
#>    67   61 0.7605555366 0.06538137 0.8253213439 0.14532774  0.77800043
#>    88   46 0.8273331347 1.00000000 0.9379403792 0.91116179  0.87088139
#>    46   83 0.9379403792 0.91116179 0.9659560170 0.85710218  1.04453191
#>    84   38 0.8905573427 0.20356443 0.9008158943 0.20198420  0.90559977
#>    38   58 0.9008158943 0.20198420 0.9470835631 0.50049843  1.22908396
#>    66   61 0.8763637907 0.16479839 0.8253213439 0.14532774  0.87843712
#>    37   98 0.6166264897 0.00000000 0.7346105442 0.03412311  0.67961061
#>    36   33 0.9370508311 0.01832039 1.0000000000 0.76296818  3.03644846
#>    66   38 0.8763637907 0.16479839 0.9008158943 0.20198420  0.89232274
#>    98   17 0.7346105442 0.03412311 0.8014866801 0.05410258  0.77255653
#>    98   67 0.7346105442 0.03412311 0.7605555366 0.06538137  0.72048067
#>    38   36 0.9008158943 0.20198420 0.9370508311 0.01832039  1.60550286
#>    66   17 0.8763637907 0.16479839 0.8014866801 0.05410258  0.88107475
#>    13   70 0.1502484818 0.63992554 0.2674387529 0.59953300  0.21727813
#>    40   78 0.1215670361 0.69023665 0.2817090879 0.67905268  0.20294058
#>    79   54 0.3162584933 0.74515231 0.4294710582 0.64113181  0.36066396
#>    54   87 0.4294710582 0.64113181 0.4111831543 0.76588825  0.39787139
#>    55   41 0.3355347859 0.79999857 0.2474956415 0.91017972  0.23267298
#>    32    1 0.4734235371 0.50549592 0.3412471555 0.57440827  0.41929741
#>    18    1 0.2211755277 0.41171380 0.3412471555 0.57440827  0.28624694
#>    18   12 0.2211755277 0.41171380 0.2283720613 0.57795693  0.28217403
#>    76   32 0.5420669327 0.59189760 0.4734235371 0.50549592  0.46942632
#>    76   54 0.5420669327 0.59189760 0.4294710582 0.64113181  0.47105774
#>    47   55 0.4078680355 0.87971615 0.3355347859 0.79999857  0.37952937
#>    29    3 0.4471462943 0.45871028 0.3769197176 0.44418620  0.39851828
#>    55   79 0.3355347859 0.79999857 0.3162584933 0.74515231  0.23029838
#>    47   87 0.4078680355 0.87971615 0.4111831543 0.76588825  0.39108944
#>    78   24 0.2817090879 0.67905268 0.1256600354 0.74130097  0.22076202
#>    99   76 0.5293945274 0.71530522 0.5420669327 0.59189760  0.50580932
#>    99   52 0.5293945274 0.71530522 0.4370319794 0.64584735  0.46333936
#>    32   53 0.4734235371 0.50549592 0.5161512630 0.49853860  0.48892552
#>    95   18 0.3682819109 0.28114786 0.2211755277 0.41171380  0.31198948
#>    12   68 0.2283720613 0.57795693 0.0463988403 0.59373594  0.13326094
#>    68   40 0.0463988403 0.59373594 0.1215670361 0.69023665  0.08880613
#>    44   24 0.0922670192 0.72170746 0.1256600354 0.74130097  0.11792900
#>    93   53 0.5388787784 0.34389511 0.5161512630 0.49853860  0.51587044
#>    94   26 0.6220334333 0.96966175 0.5513284556 0.74173575  0.54898224
#>    10   63 0.6306865955 0.67598905 0.6033538354 0.61577613  0.56915768
#>    63   16 0.6033538354 0.61577613 0.5353357860 0.49610453  0.59763706
#>    24   45 0.1256600354 0.74130097 0.1692505674 0.91081736  0.20755902
#>    59   10 0.7142486443 0.79902179 0.6306865955 0.67598905  0.63552503
#>    10   99 0.6306865955 0.67598905 0.5293945274 0.71530522  0.58002437
#>    93   29 0.5388787784 0.34389511 0.4471462943 0.45871028  0.51577886
#>    95    3 0.3682819109 0.28114786 0.3769197176 0.44418620  0.37556471
#>    15   80 0.0551180008 0.99962689 0.4165815882 0.97346585  0.24872355
#>    56   80 0.3272437655 0.84932830 0.4165815882 0.97346585  0.34169083
#>    94   48 0.6220334333 0.96966175 0.4653866597 0.77354713  0.53398183
#>    16   57 0.5353357860 0.49610453 0.6423125597 0.46359956  0.57087688
#>    30   93 0.5450761138 0.27293833 0.5388787784 0.34389511  0.53275401
#>    81   86 0.2274968009 0.26485422 0.1207797805 0.32297509  0.19767807
#>   100   62 0.5182220408 0.21540067 0.3351649278 0.25392267  0.41692195
#>    81   18 0.2274968009 0.26485422 0.2211755277 0.41171380  0.28962647
#>    35   22 0.0478217700 0.21653189 0.0000000000 0.36564842  0.04384454
#>    18   22 0.2211755277 0.41171380 0.0000000000 0.36564842  0.09466010
#>    63   72 0.6033538354 0.61577613 0.7161990718 0.50276073  0.64614284
#>    28   93 0.6809128051 0.35860639 0.5388787784 0.34389511  0.60582105
#>    28   57 0.6809128051 0.35860639 0.6423125597 0.46359956  0.60997599
#>   100   95 0.5182220408 0.21540067 0.3682819109 0.28114786  0.44860213
#>    34   81 0.2075239239 0.20126346 0.2274968009 0.26485422  0.27676795
#>    81   82 0.2274968009 0.26485422 0.1026731056 0.23936609  0.16134520
#>     5   10 0.8162889544 0.67612565 0.6306865955 0.67598905  0.72351859
#>    31    2 0.6818415372 0.26387744 0.5659857850 0.29180963  0.63184876
#>   100   51 0.5182220408 0.21540067 0.4549301015 0.34317164  0.47194095
#>    69   68 0.0469287132 0.99306009 0.0463988403 0.59373594 -1.52919555
#>    24   69 0.1256600354 0.74130097 0.0469287132 0.99306009  0.04173334
#>    45   15 0.1692505674 0.91081736 0.0551180008 0.99962689  0.20975949
#>    80   88 0.4165815882 0.97346585 0.8273331347 1.00000000  0.54239345
#>    26   97 0.5513284556 0.74173575 0.6422775611 0.96109647  0.60557266
#>    89   97 0.6677166753 0.95503520 0.6422775611 0.96109647  0.68843056
#>    97   96 0.6422775611 0.96109647 0.7029433401 0.92029182  0.61204979
#>    82   50 0.1026731056 0.23936609 0.1609054711 0.08237208  0.10700005
#>    35   86 0.0478217700 0.21653189 0.1207797805 0.32297509  0.04728174
#>    62   90 0.3351649278 0.25392267 0.5095829883 0.10875321  0.35228480
#>    65   27 0.8039568121 0.34740036 0.6869975704 0.37719243  0.75026579
#>    31   71 0.6818415372 0.26387744 0.5729773120 0.26301990  0.62718902
#>    31   60 0.6818415372 0.26387744 0.6008289234 0.20247516  0.62750414
#>    59   91 0.7142486443 0.79902179 0.8135425629 0.73571539  0.77945690
#>    28   92 0.6809128051 0.35860639 0.7263512358 0.26693038  0.70093919
#>    82   73 0.1026731056 0.23936609 0.2031979687 0.14861514  0.11862683
#>    60   11 0.6008289234 0.20247516 0.5966151683 0.16936733  0.58637257
#>    77   90 0.5537893075 0.09324029 0.5095829883 0.10875321  0.54434286
#>    65   28 0.8039568121 0.34740036 0.6809128051 0.35860639  0.74205902
#>    19    8 0.1547900962 0.05954129 0.0001700435 0.18619815  0.05501826
#>    35    8 0.0478217700 0.21653189 0.0001700435 0.18619815 -0.02345158
#>    33    5 1.0000000000 0.76296818 0.8162889544 0.67612565  0.91246068
#>    23   72 0.8692314801 0.59931976 0.7161990718 0.50276073  0.76976777
#>    96   88 0.7029433401 0.92029182 0.8273331347 1.00000000  0.74406846
#>    59   21 0.7142486443 0.79902179 0.7954210508 0.92711351  0.80474494
#>    46   21 0.9379403792 0.91116179 0.7954210508 0.92711351  0.86883659
#>    19   50 0.1547900962 0.05954129 0.1609054711 0.08237208  0.31972526
#>    85   77 0.5778672680 0.09148920 0.5537893075 0.09324029  0.56863913
#>    85   25 0.5778672680 0.09148920 0.5732349509 0.17134203  0.58166621
#>    37   19 0.6166264897 0.00000000 0.1547900962 0.05954129  0.36355898
#>    65   75 0.8039568121 0.34740036 0.8862984169 0.32953133  0.82908522
#>    60   14 0.6008289234 0.20247516 0.7047697002 0.18843387  0.65061918
#>    19   73 0.1547900962 0.05954129 0.2031979687 0.14861514  0.34168585
#>    19   74 0.1547900962 0.05954129 0.0266994967 0.19763402  0.05757085
#>     7    6 0.9032046093 0.43648666 0.7783675188 0.43791470  0.84160276
#>    43   64 0.7232782755 0.11281024 0.6826177784 0.13357991  0.71737713
#>    64   11 0.6826177784 0.13357991 0.5966151683 0.16936733  0.65079271
#>    89   94 0.6677166753 0.95503520 0.6220334333 0.96966175  0.70175350
#>    33   23 1.0000000000 0.76296818 0.8692314801 0.59931976  0.92124889
#>    33   58 1.0000000000 0.76296818 0.9470835631 0.50049843  1.00418012
#>    42    9 0.6436831113 0.05509317 0.5938670025 0.06413792  0.62638221
#>    77    9 0.5537893075 0.09324029 0.5938670025 0.06413792  0.53973465
#>    75    7 0.8862984169 0.32953133 0.9032046093 0.43648666  1.09633314
#>    84   61 0.8905573427 0.20356443 0.8253213439 0.14532774  0.80700779
#>    61   43 0.8253213439 0.14532774 0.7232782755 0.11281024  0.76356762
#>    42   85 0.6436831113 0.05509317 0.5778672680 0.09148920  0.63118274
#>    42   37 0.6436831113 0.05509317 0.6166264897 0.00000000  0.63317038
#>    77   20 0.5537893075 0.09324029 0.6023597247 0.02411119  0.49679454
#>    21    4 0.7954210508 0.92711351 0.9612747158 0.80272784  0.85694435
#>    83    4 0.9659560170 0.85710218 0.9612747158 0.80272784  0.86389612
#>    33   91 1.0000000000 0.76296818 0.8135425629 0.73571539  0.91168932
#>    37   90 0.6166264897 0.00000000 0.5095829883 0.10875321  0.38538103
#>    65   84 0.8039568121 0.34740036 0.8905573427 0.20356443  0.81693594
#>    49   46 0.9406163114 0.96377010 0.9379403792 0.91116179  0.87088139
#>    49   83 0.9406163114 0.96377010 0.9659560170 0.85710218  1.04453191
#>    49   33 0.9406163114 0.96377010 1.0000000000 0.76296818  1.84886533
#>    75   38 0.8862984169 0.32953133 0.9008158943 0.20198420  0.90559977
#>    75   58 0.8862984169 0.32953133 0.9470835631 0.50049843  1.22908396
#>    17   61 0.8014866801 0.05410258 0.8253213439 0.14532774  0.87843712
#>    42   98 0.6436831113 0.05509317 0.7346105442 0.03412311  0.67961061
#>    84   66 0.8905573427 0.20356443 0.8763637907 0.16479839  0.83263676
#>    67   17 0.7605555366 0.06538137 0.8014866801 0.05410258  0.77255653
#>    43   67 0.7232782755 0.11281024 0.7605555366 0.06538137  0.72048067
#>    58   36 0.9470835631 0.50049843 0.9370508311 0.01832039  1.60550286
#>    36   66 0.9370508311 0.01832039 0.8763637907 0.16479839  0.98126692
#>    36   17 0.9370508311 0.01832039 0.8014866801 0.05410258  0.88107475
#>    36   98 0.9370508311 0.01832039 0.7346105442 0.03412311  0.82289619
#>    36   37 0.9370508311 0.01832039 0.6166264897 0.00000000  0.80121628
#>    13   12 0.1502484818 0.63992554 0.2283720613 0.57795693  0.21727813
#>    79   78 0.3162584933 0.74515231 0.2817090879 0.67905268  0.36066396
#>    47   56 0.4078680355 0.87971615 0.3272437655 0.84932830  0.37952937
#>    29   32 0.4471462943 0.45871028 0.4734235371 0.50549592  0.39851828
#>    93   16 0.5388787784 0.34389511 0.5353357860 0.49610453  0.51587044
#>    24   41 0.1256600354 0.74130097 0.2474956415 0.91017972  0.20755902
#>    59   26 0.7142486443 0.79902179 0.5513284556 0.74173575  0.63552503
#>    94   47 0.6220334333 0.96966175 0.4078680355 0.87971615  0.53398183
#>    30    2 0.5450761138 0.27293833 0.5659857850 0.29180963  0.53275401
#>    73   62 0.2031979687 0.14861514 0.3351649278 0.25392267  0.29614455
#>     5   39 0.8162889544 0.67612565 0.7537823029 0.73139852  0.72351859
#>    31   28 0.6818415372 0.26387744 0.6809128051 0.35860639  0.63184876
#>   100   30 0.5182220408 0.21540067 0.5450761138 0.27293833  0.47194095
#>    69   22 0.0469287132 0.99306009 0.0000000000 0.36564842 -1.52919555
#>    82   35 0.1026731056 0.23936609 0.0478217700 0.21653189  0.10700005
#>    14   92 0.7047697002 0.18843387 0.7263512358 0.26693038  0.70651312
#>     7   23 0.9032046093 0.43648666 0.8692314801 0.59931976  0.84160276
#>    64   42 0.6826177784 0.13357991 0.6436831113 0.05509317  0.68314131
#>     4   33 0.9612747158 0.80272784 1.0000000000 0.76296818  1.02365448
#>    49   88 0.9406163114 0.96377010 0.8273331347 1.00000000  0.87088139
#>           my1          mx2          my2 bp1 bp2
#>   0.631245906   0.21902276  0.649261867   0   0
#>   0.524846695   0.31941447  0.631245906   0   0
#>   0.644200211   0.28319567  0.524846695   0   0
#>   0.650415262   0.31941447  0.631245906   0   0
#>   0.703295285   0.21902276  0.649261867   0   0
#>   0.679863081   0.35310662  0.650415262   0   0
#>   0.700218273   0.44662179  0.710715867   0   0
#>   0.519609889   0.28624694  0.489344670   0   0
#>   0.697083433   0.36283082  0.759594533   0   0
#>   0.808071841   0.33292465  0.939455514   0   0
#>   0.562895776   0.35310662  0.650415262   0   0
#>   0.492350556   0.28319567  0.524846695   0   0
#>   0.579139924   0.41929741  0.562895776   0   0
#>   0.582870887   0.39787139  0.700218273   0   0
#>   0.832754508   0.23267298  0.808071841   0   0
#>   0.710715867   0.43069701  0.823418795   0   0
#>   0.709376408   0.11792900  0.716224450   0   0
#>   0.516794444   0.39672960  0.519609889   0   0
#>   0.806174445   0.21977590  0.801789987   0   0
#>   0.822265266   0.36283082  0.759594533   0   0
#>   0.752987980   0.21977590  0.801789987   0   0
#>   0.650528860   0.47105774  0.582870887   0   0
#>   0.552465265   0.53172825  0.544491045   0   0
#>   0.707003868   0.44662179  0.710715867   0   0
#>   0.466017051   0.51577886  0.419492029   0   0
#>   0.365878234   0.28624694  0.489344670   0   0
#>   0.538280071   0.08880613  0.638229322   0   0
#>   0.661407203   0.11792900  0.716224450   0   0
#>   0.638229322   0.20294058  0.703295285   0   0
#>   0.419505488   0.53172825  0.544491045   0   0
#>   0.867393275   0.50763213  0.755681635   0   0
#>   0.755681635   0.46333936  0.707003868   0   0
#>   0.667609078   0.55264447  0.655338239   0   0
#>   0.498944537   0.09466010  0.465155250   0   0
#>   0.655338239   0.59763706  0.539859793   0   0
#>   0.539859793   0.53172825  0.544491045   0   0
#>   0.810603711   0.20975949  1.080619809   0   0
#>   0.939455514   0.24872355  1.164421401   0   0
#>   0.762596290   0.58002437  0.695605411   0   0
#>   0.695605411   0.50763213  0.755681635   0   0
#>   0.852134884  -0.31052427  0.793871976   0   0
#>   0.403995850   0.49724075  0.304509436   0   0
#>   0.399043358   0.49638337  0.403995850   0   0
#>   0.362510004   0.42287125  0.399043358   0   0
#>   0.933146771   0.51829227  0.916732575   0   0
#>   0.879374865   0.43069701  0.823418795   0   0
#>   0.420785878   0.60582105  0.390591443   0   0
#>   0.535293957   0.64614284  0.545655348   0   0
#>   0.307611149   0.61023903  0.347936908   0   0
#>   0.605469921   0.72353337  0.614101595   0   0
#>   0.337136567   0.09778489  0.450152080   0   0
#>   0.450152080   0.04384454  0.297482893   0   0
#>   0.188227168   0.29064579  0.341830667   0   0
#>   0.341094299   0.29064579  0.341830667   0   0
#>   0.465155250  -1.52919555  0.795489061   0   0
#>   0.545655348   0.70223266  0.439829059   0   0
#>   0.392118987   0.70223266  0.439829059   0   0
#>   0.260475588   0.37556471  0.362510004   0   0
#>   0.855954099   0.68843056  1.098385698   0   0
#>   0.733057165   0.77945690  0.791775946   0   0
#>   0.167479884   0.27676795  0.214446946   0   0
#>   0.214446946   0.28962647  0.341094299   0   0
#>   0.270424971   0.04728174  0.295126976   0   0
#>   0.634194056   0.67901701  0.733057165   0   0
#>   0.310756334   0.61023903  0.347936908   0   0
#>   0.275527467   0.54787132  0.236598227   0   0
#>   0.272036580   0.49724075  0.304509436   0   0
#>   0.249347685   0.14054451  0.180265075   0   0
#>   0.793871976   0.05905043  0.661407203   0   0
#>   0.853245190   0.09830279  0.937382480   0   0
#>   1.080619809   0.24872355  1.164421401   0   0
#>   0.937382480 -59.49839703 75.258945043   0   1
#>   2.218390931   0.70175350  1.139997342   0   0
#>   0.916732575   0.54239345  2.218390931   0   0
#>   0.847780123   0.61204979  0.850656361   0   0
#>   0.236598227   0.56141647  0.221023242   0   0
#>   0.863852002   0.74406846  0.993026674   0   0
#>   0.850656361   0.75359713  0.863852002   0   0
#>   0.890613536   0.74244003  0.995567947   0   0
#>   0.151674239   0.09550389  0.141984087   0   0
#>   0.097127284   0.34168585  0.015662093   0   0
#>   0.180265075   0.29614455  0.167479884   0   0
#>   0.527894822   0.71287832  0.437353362   0   0
#>   0.437353362   0.70223266  0.439829059   0   0
#>   0.381095727   0.71287832  0.437353362   0   0
#>   0.291428908   0.56171003  0.275527467   0   0
#>   0.251424693   0.56141647  0.221023242   0   0
#>   0.704576589   0.91246068  0.710416203   0   0
#>   0.791775946   0.78575587  0.704576589   0   0
#>   0.221023242   0.55984844  0.211001928   0   0
#>   0.211001928   0.58637257  0.187493009   0   0
#>   0.311433703   0.70651312  0.230169569   0   0
#>   0.155986856   0.31972526  0.027596742   0   0
#>   0.272574837   0.05501826  0.095448809   0   0
#>   0.187493009   0.58166621  0.131770360   0   0
#>   0.170239549   0.52025346  0.161562479   0   0
#>   0.214632690   0.70651312  0.230169569   0   0
#>   0.137063893   0.52025346  0.161562479   0   0
#>   0.348877167   0.60997599  0.392118987   0   0
#>   0.275900984  -5.59602102  0.270620530   0   1
#>   0.587408798   0.72353337  0.614101595   0   0
#>   0.406587696   0.75026579  0.381095727   0   0
#>   0.161562479   0.55984844  0.211001928   0   0
#>   0.995567947   0.70175350  1.139997342   0   0
#>   0.993026674   0.86883659  0.938399114   0   0
#>   0.831439362   0.85694435  0.836381586   0   0
#>   0.331007778   0.81693594  0.257226690   0   0
#>   0.141984087  -0.02133424  0.272574837   0   0
#>   0.131014650   0.54434286  0.137063893   0   0
#>  -0.142032087   0.34168585  0.015662093   0   0
#>   0.264541515   0.90559977  0.267128463   0   0
#>   0.179316107   0.65079271  0.178331809   0   0
#>   0.097816516   0.09550389  0.141984087   0   0
#>   0.120884056   0.58166621  0.131770360   0   0
#>   0.064924002   0.62638221  0.101513740   0   0
#>   0.508594777   0.80731937  0.527894822   0   0
#>   0.389115528   0.84043587  0.406587696   0   0
#>   0.151442723   0.65079271  0.178331809   0   0
#>   3.805466016   0.43924425  5.100441630   0   1
#>   0.691825178   0.76976777  0.587408798   0   0
#>   0.625556320   0.86079916  0.512599870   0   0
#>   0.031737975   0.61667763  0.048063453   0   0
#>   0.512599870   1.09633314  0.351145432   0   0
#>   0.231498957   0.78041103  0.209852353   0   0
#>   0.162747656   0.71737713  0.151442723   0   0
#>   0.084419736   0.69466371  0.068529676   0   0
#>   0.110194655   0.62688754  0.120884056   0   0
#>   0.048063453   0.63317038  0.026065610   0   0
#>   0.837124083   0.86389612  0.838500231   0   0
#>   0.026065610   0.38538103 -0.120553110   0   0
#>   0.001568058   0.38538103 -0.120553110   0   0
#>   0.836381586   0.91168932  0.715693694   0   0
#>   0.824746007   1.84886533  1.123187197   0   0
#>   0.068529676   0.72048067  0.072247753   0   0
#>   0.257226690   0.80700779  0.231498957   0   0
#>   0.209852353   0.76356762  0.162747656   0   0
#>   0.117456067   0.79347101  0.104923111   0   0
#>   0.940944973   0.86883659  0.938399114   0   0
#>   0.932112205   0.86389612  0.838500231   0   0
#>   0.267128463   0.89232274  0.180936674   0   0
#>   0.303947593   1.60550286  0.245605232   0   0
#>   0.082723768   0.83263676  0.202789719   0   0
#>   0.003258506   0.80121628 -0.417205250   0   0
#>   0.215831388   4.36983650  0.103112771   0   1
#>   0.180936674   0.98126692  0.122450051   0   0
#>   0.029023734   0.82289619 -0.139475310   0   0
#>   0.072247753   0.77255653  0.029023734   0   0
#>   0.245605232   0.98126692  0.122450051   0   0
#>   0.080939612   0.87843712  0.082723768   0   0
#>   0.644200211   0.21902276  0.649261867   0   0
#>   0.703295285   0.20336527  0.709376408   0   0
#>   0.679863081   0.37648614  0.697083433   0   0
#>   0.700218273   0.37648614  0.697083433   0   0
#>   0.808071841   0.23029838  0.806174445   0   0
#>   0.562895776   0.39672960  0.519609889   0   0
#>   0.489344670   0.28217403  0.492350556   0   0
#>   0.492350556   0.12985014  0.498944537   0   0
#>   0.579139924   0.50300180  0.552465265   0   0
#>   0.582870887   0.46942632  0.579139924   0   0
#>   0.832754508   0.39108944  0.822265266   0   0
#>   0.516794444   0.42287125  0.399043358   0   0
#>   0.806174445   0.36283082  0.759594533   0   0
#>   0.822265266   0.43069701  0.823418795   0   0
#>   0.752987980   0.20336527  0.709376408   0   0
#>   0.650528860   0.55264447  0.655338239   0   0
#>   0.707003868   0.50580932  0.650528860   0   0
#>   0.466017051   0.50300180  0.552465265   0   0
#>   0.365878234   0.29064579  0.341830667   0   0
#>   0.538280071   0.12985014  0.498944537   0   0
#>   0.638229322   0.05905043  0.661407203   0   0
#>   0.716224450   0.03818292  0.852134884   0   0
#>   0.419505488   0.51577886  0.419492029   0   0
#>   0.867393275   0.58585780  0.855954099   0   0
#>   0.667609078   0.70604762  0.605469921   0   0
#>   0.539859793   0.60567024  0.535293957   0   0
#>   0.810603711   0.04173334  0.853245190   0   0
#>   0.762596290   0.67901701  0.733057165   0   0
#>   0.695605411   0.56915768  0.667609078   0   0
#>   0.419492029   0.49638337  0.403995850   0   0
#>   0.362510004   0.31198948  0.365878234   0   0
#>   1.164421401   0.43986995  3.805466016   0   0
#>   0.933146771   0.33292465  0.939455514   0   0
#>   0.879374865   0.54898224  0.867393275   0   0
#>   0.420785878   0.60567024  0.535293957   0   0
#>   0.307611149   0.49724075  0.304509436   0   0
#>   0.337136567   0.16134520  0.270424971   0   0
#>   0.188227168   0.41313668  0.170239549   0   0
#>   0.341094299   0.19767807  0.337136567   0   0
#>   0.297482893  -0.02345158  0.275900984   0   0
#>   0.465155250   0.09778489  0.450152080   0   0
#>   0.545655348   0.70604762  0.605469921   0   0
#>   0.390591443   0.61023903  0.347936908   0   0
#>   0.392118987   0.60582105  0.390591443   0   0
#>   0.260475588   0.41692195  0.188227168   0   0
#>   0.214446946   0.16564903  0.249347685   0   0
#>   0.270424971   0.16564903  0.249347685   0   0
#>   0.634194056   0.72353337  0.614101595   0   0
#>   0.310756334   0.62718902  0.291428908   0   0
#>   0.272036580   0.44860213  0.260475588   0   0
#>   0.795489061  -0.31052427  0.793871976   0   0
#>   0.853245190   0.03818292  0.852134884   0   0
#>   1.080619809   0.09830279  0.937382480   0   0
#>   2.218390931   0.43986995  3.805466016   0   0
#>   0.847780123   0.58585780  0.855954099   0   0
#>   1.098385698   0.63892554  0.890613536   0   0
#>   0.850656361   0.63892554  0.890613536   0   0
#>   0.151674239   0.11862683  0.155986856   0   0
#>   0.295126976   0.04384454  0.297482893   0   0
#>   0.097127284   0.41313668  0.170239549   0   0
#>   0.381095727   0.74205902  0.348877167   0   0
#>   0.291428908   0.62750414  0.251424693   0   0
#>   0.251424693   0.65539008  0.214632690   0   0
#>   0.791775946   0.80474494  0.831439362   0   0
#>   0.311433703   0.74043160  0.331007778   0   0
#>   0.155986856   0.14054451  0.180265075   0   0
#>   0.187493009   0.65061918  0.179316107   0   0
#>   0.137063893   0.49679454  0.001568058   0   0
#>   0.348877167   0.74043160  0.331007778   0   0
#>   0.095448809  -3.11542011 -3.774956249   0   1
#>   0.275900984  -0.02133424  0.272574837   0   0
#>   0.710416203   0.92124889  0.691825178   0   0
#>   0.587408798   0.80731937  0.527894822   0   0
#>   0.993026674   0.74244003  0.995567947   0   0
#>   0.831439362   0.75359713  0.863852002   0   0
#>   0.938399114   0.85750121  0.837124083   0   0
#>   0.027596742   0.05757085  0.097816516   0   0
#>   0.131014650   0.56383264  0.064924002   0   0
#>   0.131770360   0.56863913  0.131014650   0   0
#>  -0.142032087   0.08897018 -2.271900243   0   1
#>   0.264541515   0.85611915  0.389115528   0   0
#>   0.179316107   0.65539008  0.214632690   0   0
#>   0.015662093   0.31972526  0.027596742   0   0
#>   0.097816516   0.05501826  0.095448809   0   0
#>   0.508594777   0.84043587  0.406587696   0   0
#>   0.151442723   0.68314131  0.084419736   0   0
#>   0.178331809   0.62688754  0.120884056   0   0
#>   1.139997342   0.68843056  1.098385698   0   0
#>   0.691825178   1.00418012  0.625556320   0   0
#>   0.625556320   3.03644846  0.215831388   0   0
#>   0.101513740   0.61667763  0.048063453   0   0
#>   0.031737975   0.56383264  0.064924002   0   0
#>   0.351145432   0.85611915  0.389115528   0   0
#>   0.231498957   0.83263676  0.202789719   0   0
#>   0.162747656   0.77800043  0.117456067   0   0
#>   0.110194655   0.62638221  0.101513740   0   0
#>   0.026065610   0.67961061  0.003258506   0   0
#>   0.001568058   0.53973465  0.031737975   0   0
#>   0.836381586   0.85750121  0.837124083   0   0
#>   0.838500231   1.02365448  0.824746007   0   0
#>   0.715693694   0.91246068  0.710416203   0   0
#>  -0.120553110   0.36355898 -0.142032087   0   0
#>   0.257226690   0.82908522  0.264541515   0   0
#>   0.940944973   1.04453191  0.932112205   0   0
#>   0.932112205   1.84886533  1.123187197   0   0
#>   1.123187197   6.42838271  2.477500078   0   1
#>   0.267128463   1.22908396  0.303947593   0   0
#>   0.303947593   1.09633314  0.351145432   0   0
#>   0.082723768   0.79347101  0.104923111   0   0
#>   0.003258506   0.69466371  0.068529676   0   0
#>   0.202789719   0.89232274  0.180936674   0   0
#>   0.029023734   0.79347101  0.104923111   0   0
#>   0.072247753   0.77800043  0.117456067   0   0
#>   0.245605232   3.03644846  0.215831388   0   0
#>   0.122450051   0.88107475  0.080939612   0   0
#>   0.080939612   0.82289619 -0.139475310   0   0
#>  -0.139475310   0.80121628 -0.417205250   0   0
#>  -0.417205250   0.97907137 -3.527897842   0   1
#>   0.644200211   0.13326094  0.538280071   0   0
#>   0.679863081   0.22076202  0.752987980   0   0
#>   0.832754508   0.34169083  0.933146771   0   0
#>   0.516794444   0.48892552  0.466017051   0   0
#>   0.419505488   0.57087688  0.420785878   0   0
#>   0.810603711   0.21977590  0.801789987   0   0
#>   0.762596290   0.60557266  0.847780123   0   0
#>   0.879374865   0.51829227  0.916732575   0   0
#>   0.307611149   0.56171003  0.275527467   0   0
#>   0.167479884   0.35228480  0.097127284   0   0
#>   0.634194056   0.78575587  0.704576589   0   0
#>   0.310756334   0.70093919  0.311433703   0   0
#>   0.272036580   0.54787132  0.236598227   0   0
#>   0.795489061  -3.11417801  0.914041506   0   1
#>   0.151674239   0.04728174  0.295126976   0   0
#>   0.230169569   0.78041103  0.209852353   0   0
#>   0.508594777   0.86079916  0.512599870   0   0
#>   0.084419736   0.63118274  0.110194655   0   0
#>   0.824746007   0.91168932  0.715693694   0   0
#>   0.940944973   3.43207938  8.949266154   0   1
#> 
#> $x
#>                [,1]       [,2]
#>   [1,] 0.3412471555 0.57440827
#>   [2,] 0.5659857850 0.29180963
#>   [3,] 0.3769197176 0.44418620
#>   [4,] 0.9612747158 0.80272784
#>   [5,] 0.8162889544 0.67612565
#>   [6,] 0.7783675188 0.43791470
#>   [7,] 0.9032046093 0.43648666
#>   [8,] 0.0001700435 0.18619815
#>   [9,] 0.5938670025 0.06413792
#>  [10,] 0.6306865955 0.67598905
#>  [11,] 0.5966151683 0.16936733
#>  [12,] 0.2283720613 0.57795693
#>  [13,] 0.1502484818 0.63992554
#>  [14,] 0.7047697002 0.18843387
#>  [15,] 0.0551180008 0.99962689
#>  [16,] 0.5353357860 0.49610453
#>  [17,] 0.8014866801 0.05410258
#>  [18,] 0.2211755277 0.41171380
#>  [19,] 0.1547900962 0.05954129
#>  [20,] 0.6023597247 0.02411119
#>  [21,] 0.7954210508 0.92711351
#>  [22,] 0.0000000000 0.36564842
#>  [23,] 0.8692314801 0.59931976
#>  [24,] 0.1256600354 0.74130097
#>  [25,] 0.5732349509 0.17134203
#>  [26,] 0.5513284556 0.74173575
#>  [27,] 0.6869975704 0.37719243
#>  [28,] 0.6809128051 0.35860639
#>  [29,] 0.4471462943 0.45871028
#>  [30,] 0.5450761138 0.27293833
#>  [31,] 0.6818415372 0.26387744
#>  [32,] 0.4734235371 0.50549592
#>  [33,] 1.0000000000 0.76296818
#>  [34,] 0.2075239239 0.20126346
#>  [35,] 0.0478217700 0.21653189
#>  [36,] 0.9370508311 0.01832039
#>  [37,] 0.6166264897 0.00000000
#>  [38,] 0.9008158943 0.20198420
#>  [39,] 0.7537823029 0.73139852
#>  [40,] 0.1215670361 0.69023665
#>  [41,] 0.2474956415 0.91017972
#>  [42,] 0.6436831113 0.05509317
#>  [43,] 0.7232782755 0.11281024
#>  [44,] 0.0922670192 0.72170746
#>  [45,] 0.1692505674 0.91081736
#>  [46,] 0.9379403792 0.91116179
#>  [47,] 0.4078680355 0.87971615
#>  [48,] 0.4653866597 0.77354713
#>  [49,] 0.9406163114 0.96377010
#>  [50,] 0.1609054711 0.08237208
#>  [51,] 0.4549301015 0.34317164
#>  [52,] 0.4370319794 0.64584735
#>  [53,] 0.5161512630 0.49853860
#>  [54,] 0.4294710582 0.64113181
#>  [55,] 0.3355347859 0.79999857
#>  [56,] 0.3272437655 0.84932830
#>  [57,] 0.6423125597 0.46359956
#>  [58,] 0.9470835631 0.50049843
#>  [59,] 0.7142486443 0.79902179
#>  [60,] 0.6008289234 0.20247516
#>  [61,] 0.8253213439 0.14532774
#>  [62,] 0.3351649278 0.25392267
#>  [63,] 0.6033538354 0.61577613
#>  [64,] 0.6826177784 0.13357991
#>  [65,] 0.8039568121 0.34740036
#>  [66,] 0.8763637907 0.16479839
#>  [67,] 0.7605555366 0.06538137
#>  [68,] 0.0463988403 0.59373594
#>  [69,] 0.0469287132 0.99306009
#>  [70,] 0.2674387529 0.59953300
#>  [71,] 0.5729773120 0.26301990
#>  [72,] 0.7161990718 0.50276073
#>  [73,] 0.2031979687 0.14861514
#>  [74,] 0.0266994967 0.19763402
#>  [75,] 0.8862984169 0.32953133
#>  [76,] 0.5420669327 0.59189760
#>  [77,] 0.5537893075 0.09324029
#>  [78,] 0.2817090879 0.67905268
#>  [79,] 0.3162584933 0.74515231
#>  [80,] 0.4165815882 0.97346585
#>  [81,] 0.2274968009 0.26485422
#>  [82,] 0.1026731056 0.23936609
#>  [83,] 0.9659560170 0.85710218
#>  [84,] 0.8905573427 0.20356443
#>  [85,] 0.5778672680 0.09148920
#>  [86,] 0.1207797805 0.32297509
#>  [87,] 0.4111831543 0.76588825
#>  [88,] 0.8273331347 1.00000000
#>  [89,] 0.6677166753 0.95503520
#>  [90,] 0.5095829883 0.10875321
#>  [91,] 0.8135425629 0.73571539
#>  [92,] 0.7263512358 0.26693038
#>  [93,] 0.5388787784 0.34389511
#>  [94,] 0.6220334333 0.96966175
#>  [95,] 0.3682819109 0.28114786
#>  [96,] 0.7029433401 0.92029182
#>  [97,] 0.6422775611 0.96109647
#>  [98,] 0.7346105442 0.03412311
#>  [99,] 0.5293945274 0.71530522
#> [100,] 0.5182220408 0.21540067
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,78,70), [10,2,4]
#> 2: (12,1,70), [1,3,15]
#> 3: (13,12,70), [2,10,30]
#> 4: (54,78,1), [1,13,6]
#> 5: (40,13,78), [10,20,32]
#> 6: (79,78,54), [4,9,24]
#> 7: (54,52,87), [19,9,17]
#> 8: (32,1,3), [14,21,13]
#> 9: (54,87,79), [12,6,7]
#> 10: (13,70,78), [1,5,3]
#> 11: (55,56,41), [45,22,18]
#> 12: (55,79,87), [9,23,22]
#> 13: (32,54,1), [4,8,16]
#> 14: (18,3,1), [8,15,29]
#> 15: (18,1,12), [2,40,14]
#> 16: (76,54,32), [13,26,17]
#> 17: (76,52,54), [7,16,25]
#> 18: (47,56,55), [11,23,54]
#> 19: (52,48,87), [39,7,27]
#> 20: (78,24,40), [33,5,24]
#> 21: (29,32,3), [8,51,28]
#> 22: (55,41,79), [44,12,11]
#> 23: (47,55,87), [12,39,18]
#> 24: (78,79,24), [44,20,6]
#> 25: (99,52,76), [17,41,27]
#> 26: (32,53,76), [34,16,28]
#> 27: (99,48,52), [19,25,37]
#> 28: (32,29,53), [49,26,21]
#> 29: (95,3,18), [14,63,52]
#> 30: (12,13,68), [32,40,3]
#> 31: (68,40,44), [33,85,32]
#> 32: (68,13,40), [5,31,30]
#> 33: (44,40,24), [20,48,31]
#> 34: (16,76,53), [26,35,42]
#> 35: (93,16,53), [34,49,56]
#> 36: (94,48,26), [37,72,55]
#> 37: (26,48,99), [27,47,36]
#> 38: (10,99,63), [41,59,47]
#> 39: (47,87,48), [19,55,23]
#> 40: (12,68,18), [66,15,30]
#> 41: (99,76,63), [42,38,25]
#> 42: (63,76,16), [34,57,41]
#> 43: (24,41,45), [87,86,44]
#> 44: (41,24,79), [24,22,43]
#> 45: (56,80,41), [53,11,54]
#> 46: (59,26,10), [47,73,91]
#> 47: (10,26,99), [37,38,46]
#> 48: (24,69,44), [85,33,86]
#> 49: (93,53,29), [28,50,35]
#> 50: (29,51,93), [70,49,51]
#> 51: (3,51,29), [50,21,52]
#> 52: (95,51,3), [51,29,71]
#> 53: (15,41,80), [45,146,87]
#> 54: (56,47,80), [90,45,18]
#> 55: (94,47,48), [39,36,90]
#> 56: (16,93,57), [68,57,35]
#> 57: (16,57,63), [67,42,56]
#> 58: (30,2,93), [83,70,80]
#> 59: (63,72,10), [107,38,67]
#> 60: (81,18,86), [61,76,64]
#> 61: (18,22,86), [65,60,66]
#> 62: (100,95,62), [63,115,71]
#> 63: (18,62,95), [62,29,64]
#> 64: (81,62,18), [63,60,75]
#> 65: (35,86,22), [61,121,98]
#> 66: (18,68,22), [84,61,40]
#> 67: (63,57,72), [78,59,57]
#> 68: (28,57,93), [56,83,69]
#> 69: (28,27,57), [78,68,118]
#> 70: (30,93,51), [50,81,58]
#> 71: (100,51,95), [52,62,81]
#> 72: (26,97,94), [94,36,91]
#> 73: (10,39,59), [108,46,77]
#> 74: (73,62,34), [75,100,99]
#> 75: (34,62,81), [64,82,74]
#> 76: (81,86,82), [98,82,60]
#> 77: (5,39,10), [73,107,106]
#> 78: (27,72,57), [67,69,102]
#> 79: (31,28,2), [83,104,111]
#> 80: (2,30,71), [92,104,58]
#> 81: (100,30,51), [70,71,92]
#> 82: (81,82,34), [100,75,76]
#> 83: (93,2,28), [79,68,58]
#> 84: (69,22,68), [66,85,0]
#> 85: (69,68,44), [31,48,84]
#> 86: (24,45,69), [88,48,43]
#> 87: (45,41,15), [53,88,43]
#> 88: (45,15,69), [0,86,87]
#> 89: (80,94,88), [147,146,90]
#> 90: (47,94,80), [89,54,55]
#> 91: (26,59,97), [95,72,46]
#> 92: (30,100,71), [109,80,81]
#> 93: (59,21,96), [127,95,128]
#> 94: (89,94,97), [72,96,147]
#> 95: (97,59,96), [93,96,91]
#> 96: (97,96,89), [126,94,95]
#> 97: (82,35,50), [132,112,98]
#> 98: (35,82,86), [76,65,97]
#> 99: (62,73,90), [138,115,74]
#> 100: (82,73,34), [74,82,112]
#> 101: (23,72,6), [102,142,123]
#> 102: (6,72,27), [78,103,101]
#> 103: (65,6,27), [102,118,124]
#> 104: (31,2,71), [80,105,79]
#> 105: (31,71,60), [109,116,104]
#> 106: (39,5,91), [122,108,77]
#> 107: (5,10,72), [59,123,77]
#> 108: (59,39,91), [106,128,73]
#> 109: (71,100,60), [110,105,92]
#> 110: (100,25,60), [114,109,125]
#> 111: (28,31,92), [119,130,79]
#> 112: (82,50,73), [131,100,97]
#> 113: (35,8,74), [120,132,121]
#> 114: (60,25,11), [134,137,110]
#> 115: (62,90,100), [125,62,99]
#> 116: (60,14,31), [119,105,137]
#> 117: (77,25,90), [125,161,133]
#> 118: (65,27,28), [69,130,103]
#> 119: (14,92,31), [111,116,169]
#> 120: (19,74,8), [113,0,139]
#> 121: (35,22,8), [0,113,65]
#> 122: (33,91,5), [106,148,165]
#> 123: (23,5,72), [107,101,148]
#> 124: (7,6,65), [103,143,142]
#> 125: (90,25,100), [110,115,117]
#> 126: (96,88,89), [147,96,127]
#> 127: (96,21,88), [129,126,93]
#> 128: (59,91,21), [162,93,108]
#> 129: (46,88,21), [127,159,171]
#> 130: (28,92,65), [168,118,111]
#> 131: (19,73,50), [112,139,138]
#> 132: (50,35,74), [113,139,97]
#> 133: (85,25,77), [117,141,134]
#> 134: (85,11,25), [114,133,140]
#> 135: (37,90,19), [138,0,167]
#> 136: (65,84,75), [174,143,168]
#> 137: (60,11,14), [145,116,114]
#> 138: (19,90,73), [99,131,135]
#> 139: (19,50,74), [132,120,131]
#> 140: (64,11,85), [134,157,145]
#> 141: (77,9,85), [150,133,151]
#> 142: (7,23,6), [101,124,152]
#> 143: (75,7,65), [124,136,153]
#> 144: (43,14,64), [145,156,155]
#> 145: (64,14,11), [137,140,144]
#> 146: (80,88,15), [0,53,89]
#> 147: (89,88,94), [89,94,126]
#> 148: (33,5,23), [123,149,122]
#> 149: (33,23,58), [152,178,148]
#> 150: (42,85,9), [141,158,157]
#> 151: (77,20,9), [158,141,161]
#> 152: (23,7,58), [153,149,142]
#> 153: (75,58,7), [152,143,175]
#> 154: (84,92,61), [169,179,168]
#> 155: (61,14,43), [144,170,169]
#> 156: (64,42,43), [166,144,157]
#> 157: (42,64,85), [140,150,156]
#> 158: (9,20,42), [160,150,151]
#> 159: (21,4,46), [163,129,162]
#> 160: (42,20,37), [167,177,158]
#> 161: (77,90,20), [167,151,117]
#> 162: (21,91,4), [165,159,128]
#> 163: (83,46,4), [159,164,172]
#> 164: (4,33,83), [173,163,165]
#> 165: (33,4,91), [162,122,164]
#> 166: (42,98,43), [182,156,177]
#> 167: (37,20,90), [161,135,160]
#> 168: (65,92,84), [154,136,130]
#> 169: (92,14,61), [155,154,119]
#> 170: (43,67,61), [183,155,182]
#> 171: (49,88,46), [129,172,0]
#> 172: (49,46,83), [163,173,171]
#> 173: (49,83,33), [164,0,172]
#> 174: (75,84,38), [180,175,136]
#> 175: (75,38,58), [184,153,174]
#> 176: (17,66,61), [179,183,186]
#> 177: (42,37,98), [188,166,160]
#> 178: (58,36,33), [0,149,184]
#> 179: (84,61,66), [176,180,154]
#> 180: (84,66,38), [185,174,179]
#> 181: (67,98,17), [187,183,182]
#> 182: (43,98,67), [181,170,166]
#> 183: (17,61,67), [170,181,176]
#> 184: (58,38,36), [185,178,175]
#> 185: (36,38,66), [180,186,184]
#> 186: (36,66,17), [176,187,185]
#> 187: (36,17,98), [181,188,186]
#> 188: (36,98,37), [177,0,187]
#> boundary nodes:  69 22 8 19 37 36 33 49 88 15 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.080789979 0.077967509 0.044628839 0.120396150 0.137159847 0.152550293 
#>                                                                         
#> 0.122792420 0.135019701 0.097163113 0.100312801 0.110614140 0.112930863 
#>                                                                         
#> 0.142579507 0.008910883 0.050021631 0.054741927 0.051228089 0.114332275 
#>                                                                         
#> 0.178780244 0.082983067 0.190637365 0.118080132 0.096889233 0.130809870 
#>                                                                         
#> 0.079674221 0.159093398 0.113658380 0.042998869 0.057912291 0.019338321 
#>                                                                         
#> 0.091640366 0.086539743 0.124000077 0.252346884 0.065774379 0.096029270 
#>                                                                         
#> 0.078247672 0.180541386 0.103055048 0.034346320 0.275114177 0.083951794 
#>                                                                         
#> 0.115800535 0.127630566 0.094153769 0.120748714 0.158201457 0.157084348 
#>                                                                         
#> 0.058717014 0.193184977 0.133992032 0.128096716 0.042871271 0.194657711 
#>                                                                         
#> 0.232759034 0.083623045 0.097277653 0.106559175 0.021981557 0.078331456 
#>                                                                         
#> 0.138076855 0.108221645 0.085547157 0.134991713 0.132928655 0.029611687 
#>                                                                         
#> 0.114276078 0.111559422 0.135943372 0.147399124 0.212155492 0.010497011 
#>                                                                         
#> 0.207529219 0.205487060 0.177335948 0.072565355 0.092728973 0.121795857 
#>                                                                         
#> 0.049477468 0.175461617 0.308967237 0.052825744 0.089832746 0.128919061 
#>                                                                         
#> 0.109707164 0.029626504 0.066643662 0.059652988 0.059915975 0.083611996 
#>                                                                         
#> 0.041601671 0.044614277 0.078592616 0.028889289 0.023463461 0.106996794 
#>                                                                         
#> 0.078850697 0.089268880 0.019556720 0.179450353 0.200183373 0.094061987 
#>                                                                         
#> 0.070481089 0.165828972 0.079566458 0.192254081 0.111794653 0.028342193 
#>                                                                         
#> 0.080486120 0.358189628 0.126038874 0.109822292 0.176908336 0.080102973 
#>                                                                         
#> 0.031687292 0.185223788 0.133366011 0.059157965 0.772215224 0.093284813 
#>                                                                         
#> 0.125803822 0.040917780 0.077607149 0.156787385 0.077855646 0.098319127 
#>                                                                         
#> 0.112890642 0.051647892 0.110916250 0.028015894 0.125585803 0.162220397 
#>                                                                         
#> 0.100100967 0.079498961 0.176008205 0.128026707 0.102888441 0.141866802 
#>                                                                         
#> 0.060887744 0.010379548 0.302078538 0.054630007 0.122819476 0.747303771 
#>                                                                         
#> 0.044504942 0.069796825 0.040622921 0.187204076 0.133641853 0.123956108 
#>                                                                         
#> 0.160532109 0.153744424 0.126089712 0.141034664 0.149062095 0.202204567 
#>                                                                         
#> 0.166398827 0.110350194 0.122889540 0.107642891 0.071712766 0.058135083 
#>                                                                         
#> 0.113876166 0.168006420 0.124056570 0.115564858 0.043290447 0.196692022 
#>                                                                         
#> 0.182656043 0.122321893 0.038716910 0.156304667 0.238640857 0.066126203 
#>                                                                         
#> 0.137650822 0.175031259 0.148726835 0.108654705 0.146960437 0.163266992 
#>                                                                         
#> 0.362409057 0.152942400 0.250996326 0.111806096 0.071226911 0.121517727 
#>                                                                         
#> 0.187066434 0.146995560 0.156597134 0.225921741 0.159707639 0.142793860 
#>                                                                         
#> 0.111863955 0.163721519 0.066653586 0.127399370 0.185602409 0.119175343 
#>                                                                         
#> 0.142587832 0.399324501 0.263782627 0.144614570 0.411607694 0.237467610 
#>                                                                         
#> 0.026151244 0.073111940 0.167445901 0.129046605 0.226926929 0.120693958 
#>                                                                         
#> 0.108867603 0.101652759 0.117758154 0.102318827 0.135428886 0.033374905 
#>                                                                         
#> 0.046849220 0.123553238 0.199873261 0.056487368 0.203202787 0.180949067 
#>                                                                         
#> 0.147736978 0.151645801 0.143409262 0.023635620 0.024141551 0.079987078 
#>                                                                         
#> 0.465658695 0.084258188 0.104884903 0.101377873 0.188352869 0.124845258 
#>                                                                         
#> 0.045658026 0.093151427 0.047967644 0.209478426 0.267750850 0.050630546 
#>                                                                         
#> 0.049529484 0.108283245 0.087448547 0.107098906 0.075209021 0.061378479 
#>                                                                         
#> 0.084486194 0.207314338 0.054575492 0.188438559 0.152596109 0.167894094 
#>                                                                         
#> 0.052676323 0.109636426 0.209398748 0.128370666 0.181451324 0.094287438 
#>                                                                         
#> 0.093314208 0.041282721 0.042456680 0.060324884 0.482282406 0.158551952 
#>                                                                         
#> 0.140207004 0.203056139 0.320947652 0.099716609 0.074584331 0.086160863 
#>                                                                         
#> 0.053659949 0.152250642 0.208240113 0.172698230 0.232286524 0.028166299 
#>                                                                         
#> 0.168834100 0.083439625 0.094733502 0.063495849 0.629164290 0.059414391 
#>                                                             
#> 0.081409236 0.166339388 0.087613227 0.055502054 0.118935627 
#> 
#> $alpha
#> [1] 0.2793447
#> 
#> attr(,"class")
#> [1] "scree"