Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

Usage,

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2         mx1
#>    58   60 0.425861275 0.92316963 0.556280798 0.91205904  0.49050531
#>    60   12 0.556280798 0.91205904 0.463145623 1.00000000  0.49971164
#>    12   65 0.463145623 1.00000000 0.433027626 0.96483796  0.47777683
#>    65   75 0.433027626 0.96483796 0.382212327 0.93431567  0.40534281
#>    75   19 0.382212327 0.93431567 0.410542219 0.91761089  0.40716946
#>    19   58 0.410542219 0.91761089 0.425861275 0.92316963  0.41062479
#>    70   46 0.411699965 0.01545419 0.516897102 0.10158656  0.41499794
#>    46   56 0.516897102 0.10158656 0.542080110 0.14015749  0.50741554
#>    56   69 0.542080110 0.14015749 0.537346620 0.28052054  0.56832581
#>    69   67 0.537346620 0.28052054 0.384151783 0.24675259  0.46556536
#>    67    4 0.384151783 0.24675259 0.306809300 0.18036649  0.40416212
#>     4   70 0.306809300 0.18036649 0.411699965 0.01545419  0.40316512
#>    70   43 0.411699965 0.01545419 0.708663448 0.04838830  0.59679367
#>    43   91 0.708663448 0.04838830 0.560844777 0.10081729  0.60656791
#>    91   46 0.560844777 0.10081729 0.516897102 0.10158656  0.53838646
#>    67   18 0.384151783 0.24675259 0.305763510 0.22562506  0.35360925
#>    18  100 0.305763510 0.22562506 0.147507154 0.15766835  0.22259238
#>   100   25 0.147507154 0.15766835 0.228363546 0.07689741  0.22871324
#>    25   70 0.228363546 0.07689741 0.411699965 0.01545419  0.33144129
#>    16   88 0.947296130 0.62722944 0.997825476 0.75435641  0.97509248
#>    88   92 0.997825476 0.75435641 0.878863230 0.77047147  0.94202501
#>    92   14 0.878863230 0.77047147 0.792997716 0.61286010  0.83634568
#>    14   16 0.792997716 0.61286010 0.947296130 0.62722944  0.86656012
#>    89   74 0.072177272 0.32478947 0.008685880 0.38058568  0.03468061
#>    74   59 0.008685880 0.38058568 0.000000000 0.27414299 -0.03882020
#>    59   37 0.000000000 0.27414299 0.005400755 0.25106569  0.03542889
#>    37   66 0.005400755 0.25106569 0.114040112 0.29443992  0.06496161
#>    66   89 0.114040112 0.29443992 0.072177272 0.32478947  0.06639484
#>    69   40 0.537346620 0.28052054 0.691883423 0.14909132  0.65251364
#>    40   34 0.691883423 0.14909132 0.786235911 0.21263294  0.69460895
#>    34   27 0.786235911 0.21263294 0.800037416 0.25088298  0.72358693
#>    27   84 0.800037416 0.25088298 0.830080031 0.41666381  0.83158972
#>    84   28 0.830080031 0.41666381 0.668459343 0.39691417  0.74865799
#>    28   69 0.668459343 0.39691417 0.537346620 0.28052054  0.64484649
#>    97   57 0.927872448 0.91200026 0.955975247 0.87495787  0.95859109
#>    57   32 0.955975247 0.87495787 0.972252588 0.86282010  0.96751631
#>    32   51 0.972252588 0.86282010 0.986269080 0.87475368  0.97738330
#>    51   97 0.986269080 0.87475368 0.927872448 0.91200026  1.79530785
#>    83   70 0.927872756 0.02383083 0.411699965 0.01545419  0.74730858
#>    25   55 0.228363546 0.07689741 0.112152862 0.05647842  0.17639911
#>    55   50 0.112152862 0.05647842 0.072424129 0.02481362  0.14858524
#>    54   61 0.023470855 0.86318529 0.013004167 0.75975340  0.09201780
#>    61   72 0.013004167 0.75975340 0.044556255 0.69979105  0.11124221
#>    72   81 0.044556255 0.69979105 0.086654734 0.67167053  0.11962192
#>    81   44 0.086654734 0.67167053 0.228142764 0.80996223  0.14149048
#>    44   47 0.228142764 0.80996223 0.200044559 0.93098547  0.22753495
#>    47   54 0.200044559 0.93098547 0.023470855 0.86318529  0.09885700
#>    45   21 0.091090325 0.33032253 0.232892379 0.35265365  0.15311186
#>    21   68 0.232892379 0.35265365 0.344401952 0.40104758  0.24216738
#>    68   77 0.344401952 0.40104758 0.284801181 0.63258547  0.24736543
#>    77   81 0.284801181 0.63258547 0.086654734 0.67167053  0.17240667
#>    81   87 0.086654734 0.67167053 0.056234620 0.67964583  0.04335664
#>    87   74 0.056234620 0.67964583 0.008685880 0.38058568 -0.04992439
#>    74   45 0.008685880 0.38058568 0.091090325 0.33032253  0.08149736
#>    22   65 0.245566813 0.96486930 0.433027626 0.96483796  0.33931573
#>    60   90 0.556280798 0.91205904 0.662060134 0.99735168  0.56278339
#>    34   30 0.786235911 0.21263294 0.865328732 0.12391206  0.86663622
#>    30   31 0.865328732 0.12391206 1.000000000 0.17628789  0.91624603
#>    31   23 1.000000000 0.17628789 0.978993831 0.26831155  0.93430978
#>    23   15 0.978993831 0.26831155 0.960950663 0.28079213  0.93604765
#>    15   85 0.960950663 0.28079213 0.928728101 0.46030975  0.92577699
#>    85   84 0.928728101 0.46030975 0.830080031 0.41666381  0.91001502
#>    92   64 0.878863230 0.77047147 0.648629313 0.71180726  0.76674885
#>    64   76 0.648629313 0.71180726 0.676121568 0.58007753  0.71745832
#>    76   20 0.676121568 0.58007753 0.728727757 0.56577896  0.72148898
#>    20   73 0.728727757 0.56577896 0.845047099 0.44730628  0.80702062
#>    73   39 0.845047099 0.44730628 0.905423455 0.50205408  0.82367173
#>    39   49 0.905423455 0.50205408 0.944126369 0.56592135  0.86298433
#>    49   16 0.944126369 0.56592135 0.947296130 0.62722944  0.87197754
#>    23   85 0.978993831 0.26831155 0.928728101 0.46030975  1.04935154
#>    49   88 0.944126369 0.56592135 0.997825476 0.75435641  1.24906220
#>    30   83 0.865328732 0.12391206 0.927872756 0.02383083  0.94255204
#>    67   71 0.384151783 0.24675259 0.356174514 0.31842015  0.33604243
#>    71   86 0.356174514 0.31842015 0.273463149 0.31606864  0.31587132
#>    86  100 0.273463149 0.31606864 0.147507154 0.15766835  0.20552198
#>    75   93 0.382212327 0.93431567 0.421008789 0.81173684  0.35680003
#>    93   58 0.421008789 0.81173684 0.425861275 0.92316963  0.43763508
#>    76   82 0.676121568 0.58007753 0.637010585 0.44298120  0.68363710
#>    82   28 0.637010585 0.44298120 0.668459343 0.39691417  0.72754034
#>    84   73 0.830080031 0.41666381 0.845047099 0.44730628  0.75301246
#>    86   68 0.273463149 0.31606864 0.344401952 0.40104758  0.28957178
#>    45   66 0.091090325 0.33032253 0.114040112 0.29443992  0.16071790
#>    66   86 0.114040112 0.29443992 0.273463149 0.31606864  0.19809056
#>    54   80 0.023470855 0.86318529 0.211851473 0.94813912 -1.24050211
#>    80   47 0.211851473 0.94813912 0.200044559 0.93098547  0.25328340
#>    44   75 0.228142764 0.80996223 0.382212327 0.93431567  0.29843815
#>    31   88 1.000000000 0.17628789 0.997825476 0.75435641  2.07180814
#>    49   85 0.944126369 0.56592135 0.928728101 0.46030975  1.31269708
#>    40   43 0.691883423 0.14909132 0.708663448 0.04838830  0.75173215
#>    43   83 0.708663448 0.04838830 0.927872756 0.02383083  0.81640787
#>   100   29 0.147507154 0.15766835 0.159062440 0.11311875  0.22447190
#>    29   55 0.159062440 0.11311875 0.112152862 0.05647842  0.17243520
#>    38   90 0.808829738 0.90704831 0.662060134 0.99735168  0.78401033
#>    60   64 0.556280798 0.91205904 0.648629313 0.71180726  0.64421149
#>    64   36 0.648629313 0.71180726 0.826061806 0.88534410  0.73992439
#>    36   38 0.826061806 0.88534410 0.808829738 0.90704831  0.75758419
#>    82   78 0.637010585 0.44298120 0.568719995 0.40580375  0.61814160
#>    78   52 0.568719995 0.40580375 0.531609331 0.31818362  0.61271970
#>    52   69 0.531609331 0.31818362 0.537346620 0.28052054  0.62553526
#>   100   48 0.147507154 0.15766835 0.068234296 0.08916279  0.10773062
#>    48   55 0.068234296 0.08916279 0.112152862 0.05647842  0.11302591
#>    57   92 0.955975247 0.87495787 0.878863230 0.77047147  0.92532933
#>    88   51 0.997825476 0.75435641 0.986269080 0.87475368  1.02291132
#>    77   42 0.284801181 0.63258547 0.388886508 0.57907985  0.33742020
#>    42   62 0.388886508 0.57907985 0.417958436 0.59216367  0.37624171
#>    62   63 0.417958436 0.59216367 0.438282786 0.79215955  0.40567340
#>    63   35 0.438282786 0.79215955 0.429085107 0.79523562  0.40175277
#>    35   93 0.429085107 0.79523562 0.421008789 0.81173684  0.33295656
#>    93   44 0.421008789 0.81173684 0.228142764 0.80996223  0.32500497
#>    44   77 0.228142764 0.80996223 0.284801181 0.63258547  0.22715945
#>    63   93 0.438282786 0.79215955 0.421008789 0.81173684  0.43928890
#>    97   38 0.927872448 0.91200026 0.808829738 0.90704831  0.86726030
#>    50   48 0.072424129 0.02481362 0.068234296 0.08916279 -0.32019192
#>   100   66 0.147507154 0.15766835 0.114040112 0.29443992  0.08302919
#>    97   90 0.927872448 0.91200026 0.662060134 0.99735168  0.85827164
#>    73   96 0.845047099 0.44730628 0.914887769 0.47118232  0.87436075
#>    96   85 0.914887769 0.47118232 0.928728101 0.46030975  0.94710250
#>    69   98 0.537346620 0.28052054 0.617518775 0.13641483  0.63777636
#>    98   43 0.617518775 0.13641483 0.708663448 0.04838830  0.66325524
#>    76   99 0.676121568 0.58007753 0.623688420 0.60685069  0.65671107
#>    99   53 0.623688420 0.60685069 0.594576902 0.60451670  0.61172233
#>    53   95 0.594576902 0.60451670 0.514142353 0.57610627  0.58026576
#>    95   82 0.514142353 0.57610627 0.637010585 0.44298120  0.58024370
#>    68   79 0.344401952 0.40104758 0.465407263 0.53339847  0.38491594
#>    79   62 0.465407263 0.53339847 0.417958436 0.59216367  0.42115971
#>    98   91 0.617518775 0.13641483 0.560844777 0.10081729  0.62828432
#>    91   56 0.560844777 0.10081729 0.542080110 0.14015749  0.53910534
#>    78   94 0.568719995 0.40580375 0.548470256 0.42325876  0.52316547
#>    94   68 0.548470256 0.42325876 0.344401952 0.40104758  0.44936192
#>    68   71 0.344401952 0.40104758 0.356174514 0.31842015  0.44396578
#>    99   64 0.623688420 0.60685069 0.648629313 0.71180726  0.60422327
#>    64   63 0.648629313 0.71180726 0.438282786 0.79215955  0.52601263
#>    63   95 0.438282786 0.79215955 0.514142353 0.57610627  0.51625484
#>    93   60 0.421008789 0.81173684 0.556280798 0.91205904  0.48656356
#>    74   61 0.008685880 0.38058568 0.013004167 0.75975340 -3.45485331
#>    74   72 0.008685880 0.38058568 0.044556255 0.69979105 -0.26848683
#>    79   95 0.465407263 0.53339847 0.514142353 0.57610627  0.46572646
#>    94   79 0.548470256 0.42325876 0.465407263 0.53339847  0.44436940
#>    82   94 0.637010585 0.44298120 0.548470256 0.42325876  0.58901905
#>    94   95 0.548470256 0.42325876 0.514142353 0.57610627  0.53692758
#>     1   60 0.434033256 0.94471536 0.556280798 0.91205904  0.49050531
#>     1   12 0.434033256 0.94471536 0.463145623 1.00000000  0.49971164
#>     1   65 0.434033256 0.94471536 0.433027626 0.96483796  0.47777683
#>     1   75 0.434033256 0.94471536 0.382212327 0.93431567  0.40534281
#>     1   19 0.434033256 0.94471536 0.410542219 0.91761089  0.40716946
#>     1   58 0.434033256 0.94471536 0.425861275 0.92316963  0.41062479
#>     2   46 0.504613870 0.17017680 0.516897102 0.10158656  0.41499794
#>     2   56 0.504613870 0.17017680 0.542080110 0.14015749  0.50741554
#>     2   69 0.504613870 0.17017680 0.537346620 0.28052054  0.56832581
#>     2   67 0.504613870 0.17017680 0.384151783 0.24675259  0.46556536
#>     2    4 0.504613870 0.17017680 0.306809300 0.18036649  0.40416212
#>     2   70 0.504613870 0.17017680 0.411699965 0.01545419  0.40316512
#>     3   43 0.508243790 0.05506639 0.708663448 0.04838830  0.59679367
#>     3   91 0.508243790 0.05506639 0.560844777 0.10081729  0.60656791
#>     3   46 0.508243790 0.05506639 0.516897102 0.10158656  0.53838646
#>     3   70 0.508243790 0.05506639 0.411699965 0.01545419  0.43649557
#>     4   18 0.306809300 0.18036649 0.305763510 0.22562506  0.35360925
#>     4  100 0.306809300 0.18036649 0.147507154 0.15766835  0.22259238
#>     4   25 0.306809300 0.18036649 0.228363546 0.07689741  0.22871324
#>     5   88 0.918300836 0.72800643 0.997825476 0.75435641  0.97509248
#>     5   92 0.918300836 0.72800643 0.878863230 0.77047147  0.94202501
#>     5   14 0.918300836 0.72800643 0.792997716 0.61286010  0.83634568
#>     5   16 0.918300836 0.72800643 0.947296130 0.62722944  0.86656012
#>     6   74 0.024541500 0.30420075 0.008685880 0.38058568  0.03468061
#>     6   59 0.024541500 0.30420075 0.000000000 0.27414299 -0.03882020
#>     6   37 0.024541500 0.30420075 0.005400755 0.25106569  0.03542889
#>     6   66 0.024541500 0.30420075 0.114040112 0.29443992  0.06496161
#>     6   89 0.024541500 0.30420075 0.072177272 0.32478947  0.06639484
#>     7   40 0.745691744 0.33028129 0.691883423 0.14909132  0.65251364
#>     7   34 0.745691744 0.33028129 0.786235911 0.21263294  0.69460895
#>     7   27 0.745691744 0.33028129 0.800037416 0.25088298  0.72358693
#>     7   84 0.745691744 0.33028129 0.830080031 0.41666381  0.83158972
#>     7   28 0.745691744 0.33028129 0.668459343 0.39691417  0.74865799
#>     7   69 0.745691744 0.33028129 0.537346620 0.28052054  0.64484649
#>     8   57 0.976688767 0.88061627 0.955975247 0.87495787  0.95859109
#>     8   32 0.976688767 0.88061627 0.972252588 0.86282010  0.96751631
#>     8   51 0.976688767 0.88061627 0.986269080 0.87475368  0.97738330
#>     8   97 0.976688767 0.88061627 0.927872448 0.91200026  1.79530785
#>     9   70 0.237492105 0.00000000 0.411699965 0.01545419  0.74730858
#>     9   25 0.237492105 0.00000000 0.228363546 0.07689741  0.32094379
#>     9   55 0.237492105 0.00000000 0.112152862 0.05647842  0.17639911
#>     9   50 0.237492105 0.00000000 0.072424129 0.02481362  0.14858524
#>    10   61 0.166537999 0.85546136 0.013004167 0.75975340  0.09201780
#>    10   72 0.166537999 0.85546136 0.044556255 0.69979105  0.11124221
#>    10   81 0.166537999 0.85546136 0.086654734 0.67167053  0.11962192
#>    10   44 0.166537999 0.85546136 0.228142764 0.80996223  0.14149048
#>    10   47 0.166537999 0.85546136 0.200044559 0.93098547  0.22753495
#>    10   54 0.166537999 0.85546136 0.023470855 0.86318529  0.09885700
#>    11   21 0.110631944 0.47914466 0.232892379 0.35265365  0.15311186
#>    11   68 0.110631944 0.47914466 0.344401952 0.40104758  0.24216738
#>    11   77 0.110631944 0.47914466 0.284801181 0.63258547  0.24736543
#>    11   81 0.110631944 0.47914466 0.086654734 0.67167053  0.17240667
#>    11   87 0.110631944 0.47914466 0.056234620 0.67964583  0.04335664
#>    11   74 0.110631944 0.47914466 0.008685880 0.38058568 -0.04992439
#>    11   45 0.110631944 0.47914466 0.091090325 0.33032253  0.08149736
#>    12   90 0.463145623 1.00000000 0.662060134 0.99735168  0.56278339
#>    13   34 0.903428465 0.27662054 0.786235911 0.21263294  0.86626584
#>    13   30 0.903428465 0.27662054 0.865328732 0.12391206  0.86663622
#>    13   31 0.903428465 0.27662054 1.000000000 0.17628789  0.91624603
#>    13   23 0.903428465 0.27662054 0.978993831 0.26831155  0.93430978
#>    13   15 0.903428465 0.27662054 0.960950663 0.28079213  0.93604765
#>    13   85 0.903428465 0.27662054 0.928728101 0.46030975  0.92577699
#>    13   84 0.903428465 0.27662054 0.830080031 0.41666381  0.91001502
#>    13   27 0.903428465 0.27662054 0.800037416 0.25088298  0.83521128
#>    14   64 0.792997716 0.61286010 0.648629313 0.71180726  0.76674885
#>    14   76 0.792997716 0.61286010 0.676121568 0.58007753  0.71745832
#>    14   20 0.792997716 0.61286010 0.728727757 0.56577896  0.72148898
#>    14   73 0.792997716 0.61286010 0.845047099 0.44730628  0.80702062
#>    14   39 0.792997716 0.61286010 0.905423455 0.50205408  0.82367173
#>    14   49 0.792997716 0.61286010 0.944126369 0.56592135  0.86298433
#>    17   30 0.999027192 0.04576612 0.865328732 0.12391206  0.94771031
#>    17   83 0.999027192 0.04576612 0.927872756 0.02383083  0.94255204
#>    18   71 0.305763510 0.22562506 0.356174514 0.31842015  0.33604243
#>    18   86 0.305763510 0.22562506 0.273463149 0.31606864  0.31587132
#>    19   93 0.410542219 0.91761089 0.421008789 0.81173684  0.35680003
#>    20   82 0.728727757 0.56577896 0.637010585 0.44298120  0.68363710
#>    20   28 0.728727757 0.56577896 0.668459343 0.39691417  0.72754034
#>    20   84 0.728727757 0.56577896 0.830080031 0.41666381  0.74205438
#>    21   66 0.232892379 0.35265365 0.114040112 0.29443992  0.16071790
#>    21   86 0.232892379 0.35265365 0.273463149 0.31606864  0.19809056
#>    22   80 0.245566813 0.96486930 0.211851473 0.94813912 -1.24050211
#>    22   47 0.245566813 0.96486930 0.200044559 0.93098547  0.25328340
#>    22   44 0.245566813 0.96486930 0.228142764 0.80996223  0.27067846
#>    22   75 0.245566813 0.96486930 0.382212327 0.93431567  0.29843815
#>    23   88 0.978993831 0.26831155 0.997825476 0.75435641  2.07180814
#>    23   49 0.978993831 0.26831155 0.944126369 0.56592135  1.57256707
#>    24   40 0.820780790 0.13096749 0.691883423 0.14909132  0.75809844
#>    24   43 0.820780790 0.13096749 0.708663448 0.04838830  0.75173215
#>    24   83 0.820780790 0.13096749 0.927872756 0.02383083  0.81640787
#>    24   30 0.820780790 0.13096749 0.865328732 0.12391206  0.82775683
#>    24   34 0.820780790 0.13096749 0.786235911 0.21263294  0.85342462
#>    25   29 0.228363546 0.07689741 0.159062440 0.11311875  0.22447190
#>    26   90 0.725006929 0.91918831 0.662060134 0.99735168  0.78401033
#>    26   60 0.725006929 0.91918831 0.556280798 0.91205904  0.64064170
#>    26   64 0.725006929 0.91918831 0.648629313 0.71180726  0.64421149
#>    26   36 0.725006929 0.91918831 0.826061806 0.88534410  0.73992439
#>    26   38 0.725006929 0.91918831 0.808829738 0.90704831  0.75758419
#>    28   78 0.668459343 0.39691417 0.568719995 0.40580375  0.61814160
#>    28   52 0.668459343 0.39691417 0.531609331 0.31818362  0.61271970
#>    29   48 0.159062440 0.11311875 0.068234296 0.08916279  0.10773062
#>    32   92 0.972252588 0.86282010 0.878863230 0.77047147  0.92532933
#>    32   88 0.972252588 0.86282010 0.997825476 0.75435641  0.94326509
#>    33   42 0.309227754 0.65824642 0.388886508 0.57907985  0.33742020
#>    33   62 0.309227754 0.65824642 0.417958436 0.59216367  0.37624171
#>    33   63 0.309227754 0.65824642 0.438282786 0.79215955  0.40567340
#>    33   35 0.309227754 0.65824642 0.429085107 0.79523562  0.40175277
#>    33   93 0.309227754 0.65824642 0.421008789 0.81173684  0.33295656
#>    33   44 0.309227754 0.65824642 0.228142764 0.80996223  0.32500497
#>    33   77 0.309227754 0.65824642 0.284801181 0.63258547  0.22715945
#>    36   92 0.826061806 0.88534410 0.878863230 0.77047147  0.75325651
#>    36   57 0.826061806 0.88534410 0.955975247 0.87495787  0.88815339
#>    36   97 0.826061806 0.88534410 0.927872448 0.91200026  0.88886593
#>    37   48 0.005400755 0.25106569 0.068234296 0.08916279 -0.32019192
#>    37  100 0.005400755 0.25106569 0.147507154 0.15766835  0.05980529
#>    39   96 0.905423455 0.50205408 0.914887769 0.47118232  0.87436075
#>    39   85 0.905423455 0.50205408 0.928728101 0.46030975  0.94710250
#>    40   98 0.691883423 0.14909132 0.617518775 0.13641483  0.63777636
#>    41   99 0.647220107 0.57516386 0.623688420 0.60685069  0.65671107
#>    41   53 0.647220107 0.57516386 0.594576902 0.60451670  0.61172233
#>    41   95 0.647220107 0.57516386 0.514142353 0.57610627  0.58026576
#>    41   82 0.647220107 0.57516386 0.637010585 0.44298120  0.58024370
#>    41   76 0.647220107 0.57516386 0.676121568 0.58007753  0.67374032
#>    42   68 0.388886508 0.57907985 0.344401952 0.40104758  0.28749818
#>    42   79 0.388886508 0.57907985 0.465407263 0.53339847  0.38491594
#>    45   89 0.091090325 0.33032253 0.072177272 0.32478947  0.06582727
#>    52   94 0.531609331 0.31818362 0.548470256 0.42325876  0.52316547
#>    52   68 0.531609331 0.31818362 0.344401952 0.40104758  0.44936192
#>    52   71 0.531609331 0.31818362 0.356174514 0.31842015  0.44396578
#>    52   67 0.531609331 0.31818362 0.384151783 0.24675259  0.44388257
#>    53   64 0.594576902 0.60451670 0.648629313 0.71180726  0.60422327
#>    53   63 0.594576902 0.60451670 0.438282786 0.79215955  0.52601263
#>    56   98 0.542080110 0.14015749 0.617518775 0.13641483  0.57958183
#>    59   61 0.000000000 0.27414299 0.013004167 0.75975340 -3.45485331
#>    60   63 0.556280798 0.91205904 0.438282786 0.79215955  0.49218269
#>    62   95 0.417958436 0.59216367 0.514142353 0.57610627  0.46572646
#>    72   87 0.044556255 0.69979105 0.056234620 0.67964583 -0.16952291
#>    73   85 0.845047099 0.44730628 0.928728101 0.46030975  0.89460849
#>     9   83 0.237492105 0.00000000 0.927872756 0.02383083  0.74730858
#>    12   22 0.463145623 1.00000000 0.245566813 0.96486930  0.33931573
#>    17   31 0.999027192 0.04576612 1.000000000 0.17628789  0.94771031
#>    22   54 0.245566813 0.96486930 0.023470855 0.86318529 -1.24050211
#>    37   50 0.005400755 0.25106569 0.072424129 0.02481362 -0.32019192
#>           my1          mx2           my2 bp1 bp2
#>   0.910973710   0.48656356   0.864704222   0   0
#>   0.945437213   0.56278339   1.012234286   0   0
#>   0.956987879   0.33931573   1.075586529   0   0
#>   0.953367977   0.33931368   1.063297152   0   0
#>   0.944265910   0.35680003   0.858843628   0   0
#>   0.941271227   0.43763508   0.866834876   0   0
#>   0.118733292   0.43649557   0.092477318   0   0
#>   0.135283560   0.53910534   0.114593245   0   0
#>   0.211303917   0.58344729   0.211813861   0   0
#>   0.241787134   0.45552759   0.287325413   0   0
#>   0.145193126   0.35360925   0.204089264   0   0
#>   0.125839125   0.33144129   0.080219966   0   0
#>  -0.298205136   0.68834052  -1.123673706   0   0
#>  -0.004865575   0.62828432   0.056361829   0   0
#>   0.073524387   0.53910534   0.114593245   0   0
#>   0.204089264   0.33604243   0.269266471   0   0
#>   0.201061857   0.20552198   0.240815084   0   0
#>   0.158103904   0.22447190   0.153858075   0   0
#>   0.080219966   0.32094379   0.048897156   0   0
#>   0.689786653   1.24906220   0.580891497   0   0
#>   0.789584771   0.94326509   0.798739057   0   0
#>   0.691439580   0.76674885   0.729355419   0   0
#>   0.658560001   0.87197754   0.600387589   0   0
#>   0.346143460   0.06582727   0.381585742   0   0
#>   0.330886511  -3.45485331   0.609639874   0   0
#>   0.270263759 -41.04711713  -9.344236697   0   1
#>   0.259625244   0.08302919   0.214371404   0   0
#>   0.272766779   0.08867266   0.303495823   0   0
#>   0.259367799   0.63777636   0.242039444   0   0
#>   0.246866679   0.75809844   0.152591599   0   0
#>   0.256853119   0.86626584   0.205371255   0   0
#>   0.330777668   0.83521128   0.330121373   0   0
#>   0.411794835   0.74205438   0.465835249   0   0
#>   0.291469687   0.62553526   0.313223009   0   0
#>   0.906123935   0.88886593   0.853225838   0   0
#>   0.873451756   0.92532933   0.816876938   0   0
#>   0.870992140   1.02291132   0.817517546   0   0
#>   2.207597576   9.55909714  14.379967410   0   1
#>  -4.757314116   0.68834052  -1.123673706   0   0
#>   0.031738129   0.17243520   0.054297972   0   0
#>  -0.029987568   0.07882260   0.057541217   0   0
#>   0.804003226  -9.47817931   1.772449920   0   1
#>   0.773163596  -0.26848683   0.573350810   0   0
#>   0.766597335   0.07956402   0.706627725   0   0
#>   0.757092333   0.19552724   0.701806633   0   0
#>   0.873594542   0.27067846   0.883611257   0   0
#>   0.930683006   0.05575185   1.042942836   0   0
#>   0.397872628   0.16071790   0.349574496   0   0
#>   0.483949627   0.28957178   0.374720116   0   0
#>   0.499509097   0.28749818   0.509839772   0   0
#>   0.584594097   0.19552724   0.701806633   0   0
#>   0.568522181   0.07956402   0.706627725   0   0
#>   0.543214406  -0.16952291   0.562229841   0   0
#>   0.407276225   0.06582727   0.381585742   0   0
#>   1.075586529   0.33931368   1.063297152   0   0
#>   1.012234286   0.64064170   0.915674958   0   0
#>   0.204692902   0.85342462   0.192915033   0   0
#>   0.192315589   0.94771031   0.111413106   0   0
#>   0.209702195   2.07180814   0.469358066   0   0
#>   0.225507097   1.04935154   0.389310378   0   0
#>   0.367129333   1.04935154   0.389310378   0   0
#>   0.369300240   0.89460849   0.404121908   0   0
#>   0.729355419   0.75325651   0.782307551   0   0
#>   0.657438303   0.67840993   0.649288826   0   0
#>   0.643068226   0.68363710   0.503806514   0   0
#>   0.526309881   0.75301246   0.473283360   0   0
#>   0.531544915   0.87436075   0.475644615   0   0
#>   0.571432172   0.96643877   0.508739842   0   0
#>   0.600387589   1.24906220   0.580891497   0   0
#>   0.389310378   1.31269708   0.458255064   0   0
#>   0.580891497   1.57256707   0.488700994   0   0
#>   0.102587911   0.82775683   0.030848638   0   0
#>   0.269266471   0.44388257   0.311364632   0   0
#>   0.280224440   0.31375881   0.354529239   0   0
#>   0.240815084   0.20212786   0.243514011   0   0
#>   0.858843628   0.32422902   0.848534834   0   0
#>   0.866834876   0.48656356   0.864704222   0   0
#>   0.503806514   0.67374032   0.506629879   0   0
#>   0.471015361   0.61814160   0.396331696   0   0
#>   0.473283360   0.89460849   0.404121908   0   0
#>   0.374720116   0.31375881   0.354529239   0   0
#>   0.349574496   0.08867266   0.303495823   0   0
#>   0.273272465   0.20212786   0.243514011   0   0
#>   3.917317638   0.05575185   1.042942836   0   0
#>   0.906981185   0.05575185   1.042942836   0   0
#>   0.880488830   0.32422902   0.848534834   0   0
#>   0.469358066   3.80168241   0.475865345   0   1
#>   0.458255064   0.96643877   0.508739842   0   0
#>   0.107314311   0.66325524   0.092571510   0   0
#>   0.019504414   0.68834052  -1.123673706   0   0
#>   0.153858075   0.10773062   0.123577696   0   0
#>   0.054297972   0.11302591   0.103500804   0   0
#>   1.031133080   0.85827164   1.151829601   0   0
#>   0.831189634   0.55863782   0.791726308   0   0
#>   0.795938953   0.75325651   0.782307551   0   0
#>   0.848669071   0.86726030   0.935746471   0   0
#>   0.396331696   0.58901905   0.449826377   0   0
#>   0.335499098   0.52316547   0.373428933   0   0
#>   0.313223009   0.45552759   0.287325413   0   0
#>   0.123577696   0.05980529   0.179035654   0   0
#>   0.103500804   0.07882260   0.057541217   0   0
#>   0.816876938   0.88815339   0.844313161   0   0
#>   0.817517546   9.25292209   1.607479482   0   1
#>   0.606953852   0.28749818   0.509839772   0   0
#>   0.646016720   0.42115971   0.546210061   0   0
#>   0.694442781   0.48315125   0.686569184   0   0
#>   0.698221173   0.43928890   0.810456779   0   0
#>   0.758413741   0.43928890   0.810456779   0   0
#>   0.764204569   0.32422902   0.848534834   0   0
#>   0.711910724   0.19552724   0.701806633   0   0
#>   0.810456779   0.49218269   0.857127528   0   0
#>   0.935746471   0.85827164   1.151829601   0   0
#>   0.031561023   0.07882260   0.057541217   0   0
#>   0.214371404   0.20212786   0.243514011   0   0
#>   1.151829601   1.95334818   4.562257261   0   1
#>   0.475644615   0.89110570   0.426663372   0   0
#>   0.497944960   0.89110570   0.426663372   0   0
#>   0.242039444   0.58344729   0.211813861   0   0
#>   0.092571510   0.62828432   0.056361829   0   0
#>   0.606793280   0.67840993   0.649288826   0   0
#>   0.573383150   0.60422327   0.666917847   0   0
#>   0.516966991   0.51625484   0.698192354   0   0
#>   0.513851370   0.57569300   0.509651288   0   0
#>   0.485498198   0.44436940   0.431141294   0   0
#>   0.546210061   0.46572646   0.582194626   0   0
#>   0.056361829   0.57958183   0.133899904   0   0
#>   0.114593245   0.57958183   0.133899904   0   0
#>   0.373428933   0.58901905   0.449826377   0   0
#>   0.385271849   0.44436940   0.431141294   0   0
#>   0.373080822   0.31375881   0.354529239   0   0
#>   0.666917847   0.67840993   0.649288826   0   0
#>   0.706319949   0.55863782   0.791726308   0   0
#>   0.698192354   0.48315125   0.686569184   0   0
#>   0.864704222   0.49218269   0.857127528   0   0
#>   0.609639874  -0.26848683   0.573350810   0   0
#>   0.573350810  -0.16952291   0.562229841   0   0
#>   0.582194626   0.53692758   0.500944993   0   0
#>   0.431141294   0.53692758   0.500944993   0   0
#>   0.449826377   0.57569300   0.509651288   0   0
#>   0.500944993   0.57569300   0.509651288   0   0
#>   0.910973710   0.49971164   0.945437213   0   0
#>   0.945437213   0.47777683   0.956987879   0   0
#>   0.956987879   0.40534281   0.953367977   0   0
#>   0.953367977   0.40716946   0.944265910   0   0
#>   0.944265910   0.41062479   0.941271227   0   0
#>   0.941271227   0.49050531   0.910973710   0   0
#>   0.118733292   0.50741554   0.135283560   0   0
#>   0.135283560   0.56832581   0.211303917   0   0
#>   0.211303917   0.46556536   0.241787134   0   0
#>   0.241787134   0.40416212   0.145193126   0   0
#>   0.145193126   0.40316512   0.125839125   0   0
#>   0.125839125   0.41499794   0.118733292   0   0
#>  -0.298205136   0.60656791  -0.004865575   0   0
#>  -0.004865575   0.53838646   0.073524387   0   0
#>   0.073524387   0.43649557   0.092477318   0   0
#>   0.092477318   0.59679367  -0.298205136   0   0
#>   0.204089264   0.22259238   0.201061857   0   0
#>   0.201061857   0.22871324   0.158103904   0   0
#>   0.158103904   0.33144129   0.080219966   0   0
#>   0.689786653   0.94202501   0.789584771   0   0
#>   0.789584771   0.83634568   0.691439580   0   0
#>   0.691439580   0.86656012   0.658560001   0   0
#>   0.658560001   0.97509248   0.689786653   0   0
#>   0.346143460  -0.03882020   0.330886511   0   0
#>   0.330886511   0.03542889   0.270263759   0   0
#>   0.270263759   0.06496161   0.259625244   0   0
#>   0.259625244   0.06639484   0.272766779   0   0
#>   0.272766779   0.03468061   0.346143460   0   0
#>   0.259367799   0.69460895   0.246866679   0   0
#>   0.246866679   0.72358693   0.256853119   0   0
#>   0.256853119   0.83158972   0.330777668   0   0
#>   0.330777668   0.74865799   0.411794835   0   0
#>   0.411794835   0.64484649   0.291469687   0   0
#>   0.291469687   0.65251364   0.259367799   0   0
#>   0.906123935   0.96751631   0.873451756   0   0
#>   0.873451756   0.97738330   0.870992140   0   0
#>   0.870992140   1.79530785   2.207597576   0   0
#>   2.207597576   0.95859109   0.906123935   0   0
#>  -4.757314116   0.32094379   0.048897156   0   0
#>   0.048897156   0.17639911   0.031738129   0   0
#>   0.031738129   0.14858524  -0.029987568   0   0
#>  -0.029987568  -0.74196785  -5.954226640   0   1
#>   0.804003226   0.11124221   0.773163596   0   0
#>   0.773163596   0.11962192   0.766597335   0   0
#>   0.766597335   0.14149048   0.757092333   0   0
#>   0.757092333   0.22753495   0.873594542   0   0
#>   0.873594542   0.09885700   0.930683006   0   0
#>   0.930683006   0.09201780   0.804003226   0   0
#>   0.397872628   0.24216738   0.483949627   0   0
#>   0.483949627   0.24736543   0.499509097   0   0
#>   0.499509097   0.17240667   0.584594097   0   0
#>   0.584594097   0.04335664   0.568522181   0   0
#>   0.568522181  -0.04992439   0.543214406   0   0
#>   0.543214406   0.08149736   0.407276225   0   0
#>   0.407276225   0.15311186   0.397872628   0   0
#>   1.012234286   0.62970416   6.038628640   0   1
#>   0.205371255   0.86663622   0.204692902   0   0
#>   0.204692902   0.91624603   0.192315589   0   0
#>   0.192315589   0.93430978   0.209702195   0   0
#>   0.209702195   0.93604765   0.225507097   0   0
#>   0.225507097   0.92577699   0.367129333   0   0
#>   0.367129333   0.91001502   0.369300240   0   0
#>   0.369300240   0.83521128   0.330121373   0   0
#>   0.330121373   0.86626584   0.205371255   0   0
#>   0.729355419   0.71745832   0.657438303   0   0
#>   0.657438303   0.72148898   0.643068226   0   0
#>   0.643068226   0.80702062   0.526309881   0   0
#>   0.526309881   0.82367173   0.531544915   0   0
#>   0.531544915   0.86298433   0.571432172   0   0
#>   0.571432172   0.87197754   0.600387589   0   0
#>   0.111413106   0.94255204   0.102587911   0   0
#>   0.102587911   4.89905424 -12.731650457   0   1
#>   0.269266471   0.31587132   0.280224440   0   0
#>   0.280224440   0.20552198   0.240815084   0   0
#>   0.858843628   0.43763508   0.866834876   0   0
#>   0.503806514   0.72754034   0.471015361   0   0
#>   0.471015361   0.74205438   0.465835249   0   0
#>   0.465835249   0.75301246   0.473283360   0   0
#>   0.349574496   0.19809056   0.273272465   0   0
#>   0.273272465   0.28957178   0.374720116   0   0
#>   3.917317638   0.25328340   0.906981185   0   0
#>   0.906981185   0.27067846   0.883611257   0   0
#>   0.883611257   0.29843815   0.880488830   0   0
#>   0.880488830   0.33931368   1.063297152   0   0
#>   0.469358066   1.57256707   0.488700994   0   0
#>   0.488700994   1.31269708   0.458255064   0   0
#>   0.152591599   0.75173215   0.107314311   0   0
#>   0.107314311   0.81640787   0.019504414   0   0
#>   0.019504414   0.82775683   0.030848638   0   0
#>   0.030848638   0.85342462   0.192915033   0   0
#>   0.192915033   0.75809844   0.152591599   0   0
#>   0.153858075   0.17243520   0.054297972   0   0
#>   1.031133080   0.64064170   0.915674958   0   0
#>   0.915674958   0.64421149   0.831189634   0   0
#>   0.831189634   0.73992439   0.795938953   0   0
#>   0.795938953   0.75758419   0.848669071   0   0
#>   0.848669071   0.78401033   1.031133080   0   0
#>   0.396331696   0.61271970   0.335499098   0   0
#>   0.335499098   0.62553526   0.313223009   0   0
#>   0.123577696   0.11302591   0.103500804   0   0
#>   0.816876938   0.94326509   0.798739057   0   0
#>   0.798739057   1.02291132   0.817517546   0   0
#>   0.606953852   0.37624171   0.646016720   0   0
#>   0.646016720   0.40567340   0.694442781   0   0
#>   0.694442781   0.40175277   0.698221173   0   0
#>   0.698221173   0.33295656   0.758413741   0   0
#>   0.758413741   0.32500497   0.764204569   0   0
#>   0.764204569   0.22715945   0.711910724   0   0
#>   0.711910724   0.33742020   0.606953852   0   0
#>   0.782307551   0.88815339   0.844313161   0   0
#>   0.844313161   0.88886593   0.853225838   0   0
#>   0.853225838   0.86726030   0.935746471   0   0
#>   0.031561023   0.05980529   0.179035654   0   0
#>   0.179035654   0.08302919   0.214371404   0   0
#>   0.475644615   0.94710250   0.497944960   0   0
#>   0.497944960   0.96643877   0.508739842   0   0
#>   0.242039444   0.66325524   0.092571510   0   0
#>   0.606793280   0.61172233   0.573383150   0   0
#>   0.573383150   0.58026576   0.516966991   0   0
#>   0.516966991   0.58024370   0.513851370   0   0
#>   0.513851370   0.67374032   0.506629879   0   0
#>   0.506629879   0.65671107   0.606793280   0   0
#>   0.509839772   0.38491594   0.485498198   0   0
#>   0.485498198   0.42115971   0.546210061   0   0
#>   0.381585742   0.08867266   0.303495823   0   0
#>   0.373428933   0.44936192   0.385271849   0   0
#>   0.385271849   0.44396578   0.373080822   0   0
#>   0.373080822   0.44388257   0.311364632   0   0
#>   0.311364632   0.45552759   0.287325413   0   0
#>   0.666917847   0.52601263   0.706319949   0   0
#>   0.706319949   0.51625484   0.698192354   0   0
#>   0.133899904   0.58344729   0.211813861   0   0
#>   0.609639874  -5.51264157   0.664745408   0   1
#>   0.857127528   0.55863782   0.791726308   0   0
#>   0.582194626   0.48315125   0.686569184   0   0
#>   0.562229841   0.07956402   0.706627725   0   0
#>   0.404121908   0.89110570   0.426663372   0   0
#>  -4.757314116   0.79724815  -6.204066567   0   1
#>   1.075586529  -0.38391459   5.554848132   0   1
#>   0.111413106   8.60884163   0.054313006   0   1
#>   3.917317638  -2.94471519   7.639622157   0   1
#>   0.031561023  -4.38347146  -1.172117244   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.434033256 0.94471536
#>   [2,] 0.504613870 0.17017680
#>   [3,] 0.508243790 0.05506639
#>   [4,] 0.306809300 0.18036649
#>   [5,] 0.918300836 0.72800643
#>   [6,] 0.024541500 0.30420075
#>   [7,] 0.745691744 0.33028129
#>   [8,] 0.976688767 0.88061627
#>   [9,] 0.237492105 0.00000000
#>  [10,] 0.166537999 0.85546136
#>  [11,] 0.110631944 0.47914466
#>  [12,] 0.463145623 1.00000000
#>  [13,] 0.903428465 0.27662054
#>  [14,] 0.792997716 0.61286010
#>  [15,] 0.960950663 0.28079213
#>  [16,] 0.947296130 0.62722944
#>  [17,] 0.999027192 0.04576612
#>  [18,] 0.305763510 0.22562506
#>  [19,] 0.410542219 0.91761089
#>  [20,] 0.728727757 0.56577896
#>  [21,] 0.232892379 0.35265365
#>  [22,] 0.245566813 0.96486930
#>  [23,] 0.978993831 0.26831155
#>  [24,] 0.820780790 0.13096749
#>  [25,] 0.228363546 0.07689741
#>  [26,] 0.725006929 0.91918831
#>  [27,] 0.800037416 0.25088298
#>  [28,] 0.668459343 0.39691417
#>  [29,] 0.159062440 0.11311875
#>  [30,] 0.865328732 0.12391206
#>  [31,] 1.000000000 0.17628789
#>  [32,] 0.972252588 0.86282010
#>  [33,] 0.309227754 0.65824642
#>  [34,] 0.786235911 0.21263294
#>  [35,] 0.429085107 0.79523562
#>  [36,] 0.826061806 0.88534410
#>  [37,] 0.005400755 0.25106569
#>  [38,] 0.808829738 0.90704831
#>  [39,] 0.905423455 0.50205408
#>  [40,] 0.691883423 0.14909132
#>  [41,] 0.647220107 0.57516386
#>  [42,] 0.388886508 0.57907985
#>  [43,] 0.708663448 0.04838830
#>  [44,] 0.228142764 0.80996223
#>  [45,] 0.091090325 0.33032253
#>  [46,] 0.516897102 0.10158656
#>  [47,] 0.200044559 0.93098547
#>  [48,] 0.068234296 0.08916279
#>  [49,] 0.944126369 0.56592135
#>  [50,] 0.072424129 0.02481362
#>  [51,] 0.986269080 0.87475368
#>  [52,] 0.531609331 0.31818362
#>  [53,] 0.594576902 0.60451670
#>  [54,] 0.023470855 0.86318529
#>  [55,] 0.112152862 0.05647842
#>  [56,] 0.542080110 0.14015749
#>  [57,] 0.955975247 0.87495787
#>  [58,] 0.425861275 0.92316963
#>  [59,] 0.000000000 0.27414299
#>  [60,] 0.556280798 0.91205904
#>  [61,] 0.013004167 0.75975340
#>  [62,] 0.417958436 0.59216367
#>  [63,] 0.438282786 0.79215955
#>  [64,] 0.648629313 0.71180726
#>  [65,] 0.433027626 0.96483796
#>  [66,] 0.114040112 0.29443992
#>  [67,] 0.384151783 0.24675259
#>  [68,] 0.344401952 0.40104758
#>  [69,] 0.537346620 0.28052054
#>  [70,] 0.411699965 0.01545419
#>  [71,] 0.356174514 0.31842015
#>  [72,] 0.044556255 0.69979105
#>  [73,] 0.845047099 0.44730628
#>  [74,] 0.008685880 0.38058568
#>  [75,] 0.382212327 0.93431567
#>  [76,] 0.676121568 0.58007753
#>  [77,] 0.284801181 0.63258547
#>  [78,] 0.568719995 0.40580375
#>  [79,] 0.465407263 0.53339847
#>  [80,] 0.211851473 0.94813912
#>  [81,] 0.086654734 0.67167053
#>  [82,] 0.637010585 0.44298120
#>  [83,] 0.927872756 0.02383083
#>  [84,] 0.830080031 0.41666381
#>  [85,] 0.928728101 0.46030975
#>  [86,] 0.273463149 0.31606864
#>  [87,] 0.056234620 0.67964583
#>  [88,] 0.997825476 0.75435641
#>  [89,] 0.072177272 0.32478947
#>  [90,] 0.662060134 0.99735168
#>  [91,] 0.560844777 0.10081729
#>  [92,] 0.878863230 0.77047147
#>  [93,] 0.421008789 0.81173684
#>  [94,] 0.548470256 0.42325876
#>  [95,] 0.514142353 0.57610627
#>  [96,] 0.914887769 0.47118232
#>  [97,] 0.927872448 0.91200026
#>  [98,] 0.617518775 0.13641483
#>  [99,] 0.623688420 0.60685069
#> [100,] 0.147507154 0.15766835
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.4340333,0.9447154) [6]: 12 19 58 60 65 75 
#> 2: (0.5046139,0.1701768) [6]: 4 46 56 67 69 70 
#> 3: (0.5082438,0.05506639) [4]: 43 46 70 91 
#> 4: (0.3068093,0.1803665) [6]: 2 18 25 67 70 100 
#> 5: (0.9183008,0.7280064) [4]: 14 16 88 92 
#> 6: (0.0245415,0.3042008) [5]: 37 59 66 74 89 
#> 7: (0.7456917,0.3302813) [6]: 27 28 34 40 69 84 
#> 8: (0.9766888,0.8806163) [4]: 32 51 57 97 
#> 9: (0.2374921,0) [5]: 25 50 55 70 83 
#> 10: (0.166538,0.8554614) [6]: 44 47 54 61 72 81 
#> 11: (0.1106319,0.4791447) [7]: 21 45 68 74 77 81 87 
#> 12: (0.4631456,1) [5]: 1 22 60 65 90 
#> 13: (0.9034285,0.2766205) [8]: 15 23 27 30 31 34 84 85 
#> 14: (0.7929977,0.6128601) [9]: 5 16 20 39 49 64 73 76 92 
#> 15: (0.9609507,0.2807921) [3]: 13 23 85 
#> 16: (0.9472961,0.6272294) [4]: 5 14 49 88 
#> 17: (0.9990272,0.04576612) [3]: 30 31 83 
#> 18: (0.3057635,0.2256251) [5]: 4 67 71 86 100 
#> 19: (0.4105422,0.9176109) [4]: 1 58 75 93 
#> 20: (0.7287278,0.565779) [6]: 14 28 73 76 82 84 
#> 21: (0.2328924,0.3526537) [5]: 11 45 66 68 86 
#> 22: (0.2455668,0.9648693) [7]: 12 44 47 54 65 75 80 
#> 23: (0.9789938,0.2683115) [6]: 13 15 31 49 85 88 
#> 24: (0.8207808,0.1309675) [5]: 30 34 40 43 83 
#> 25: (0.2283635,0.07689741) [6]: 4 9 29 55 70 100 
#> 26: (0.7250069,0.9191883) [5]: 36 38 60 64 90 
#> 27: (0.8000374,0.250883) [4]: 7 13 34 84 
#> 28: (0.6684593,0.3969142) [7]: 7 20 52 69 78 82 84 
#> 29: (0.1590624,0.1131187) [4]: 25 48 55 100 
#> 30: (0.8653287,0.1239121) [6]: 13 17 24 31 34 83 
#> 31: (1,0.1762879) [5]: 13 17 23 30 88 
#> 32: (0.9722526,0.8628201) [5]: 8 51 57 88 92 
#> 33: (0.3092278,0.6582464) [7]: 35 42 44 62 63 77 93 
#> 34: (0.7862359,0.2126329) [6]: 7 13 24 27 30 40 
#> 35: (0.4290851,0.7952356) [3]: 33 63 93 
#> 36: (0.8260618,0.8853441) [6]: 26 38 57 64 92 97 
#> 37: (0.005400755,0.2510657) [6]: 6 48 50 59 66 100 
#> 38: (0.8088297,0.9070483) [4]: 26 36 90 97 
#> 39: (0.9054235,0.5020541) [5]: 14 49 73 85 96 
#> 40: (0.6918834,0.1490913) [6]: 7 24 34 43 69 98 
#> 41: (0.6472201,0.5751639) [5]: 53 76 82 95 99 
#> 42: (0.3888865,0.5790798) [5]: 33 62 68 77 79 
#> 43: (0.7086634,0.0483883) [7]: 3 24 40 70 83 91 98 
#> 44: (0.2281428,0.8099622) [8]: 10 22 33 47 75 77 81 93 
#> 45: (0.09109033,0.3303225) [5]: 11 21 66 74 89 
#> 46: (0.5168971,0.1015866) [5]: 2 3 56 70 91 
#> 47: (0.2000446,0.9309855) [5]: 10 22 44 54 80 
#> 48: (0.0682343,0.08916279) [5]: 29 37 50 55 100 
#> 49: (0.9441264,0.5659214) [6]: 14 16 23 39 85 88 
#> 50: (0.07242413,0.02481362) [4]: 9 37 48 55 
#> 51: (0.9862691,0.8747537) [4]: 8 32 88 97 
#> 52: (0.5316093,0.3181836) [7]: 28 67 68 69 71 78 94 
#> 53: (0.5945769,0.6045167) [5]: 41 63 64 95 99 
#> 54: (0.02347085,0.8631853) [5]: 10 22 47 61 80 
#> 55: (0.1121529,0.05647842) [5]: 9 25 29 48 50 
#> 56: (0.5420801,0.1401575) [5]: 2 46 69 91 98 
#> 57: (0.9559752,0.8749579) [5]: 8 32 36 92 97 
#> 58: (0.4258613,0.9231696) [4]: 1 19 60 93 
#> 59: (0,0.274143) [4]: 6 37 61 74 
#> 60: (0.5562808,0.912059) [8]: 1 12 26 58 63 64 90 93 
#> 61: (0.01300417,0.7597534) [5]: 10 54 59 72 74 
#> 62: (0.4179584,0.5921637) [5]: 33 42 63 79 95 
#> 63: (0.4382828,0.7921595) [8]: 33 35 53 60 62 64 93 95 
#> 64: (0.6486293,0.7118073) [9]: 14 26 36 53 60 63 76 92 99 
#> 65: (0.4330276,0.964838) [4]: 1 12 22 75 
#> 66: (0.1140401,0.2944399) [7]: 6 21 37 45 86 89 100 
#> 67: (0.3841518,0.2467526) [6]: 2 4 18 52 69 71 
#> 68: (0.344402,0.4010476) [9]: 11 21 42 52 71 77 79 86 94 
#> 69: (0.5373466,0.2805205) [8]: 2 7 28 40 52 56 67 98 
#> 70: (0.4117,0.01545419) [8]: 2 3 4 9 25 43 46 83 
#> 71: (0.3561745,0.3184202) [5]: 18 52 67 68 86 
#> 72: (0.04455625,0.6997911) [5]: 10 61 74 81 87 
#> 73: (0.8450471,0.4473063) [6]: 14 20 39 84 85 96 
#> 74: (0.00868588,0.3805857) [8]: 6 11 45 59 61 72 87 89 
#> 75: (0.3822123,0.9343157) [6]: 1 19 22 44 65 93 
#> 76: (0.6761216,0.5800775) [6]: 14 20 41 64 82 99 
#> 77: (0.2848012,0.6325855) [6]: 11 33 42 44 68 81 
#> 78: (0.56872,0.4058038) [4]: 28 52 82 94 
#> 79: (0.4654073,0.5333985) [5]: 42 62 68 94 95 
#> 80: (0.2118515,0.9481391) [3]: 22 47 54 
#> 81: (0.08665473,0.6716705) [6]: 10 11 44 72 77 87 
#> 82: (0.6370106,0.4429812) [7]: 20 28 41 76 78 94 95 
#> 83: (0.9278728,0.02383083) [6]: 9 17 24 30 43 70 
#> 84: (0.83008,0.4166638) [7]: 7 13 20 27 28 73 85 
#> 85: (0.9287281,0.4603098) [8]: 13 15 23 39 49 73 84 96 
#> 86: (0.2734631,0.3160686) [6]: 18 21 66 68 71 100 
#> 87: (0.05623462,0.6796458) [4]: 11 72 74 81 
#> 88: (0.9978255,0.7543564) [8]: 5 16 23 31 32 49 51 92 
#> 89: (0.07217727,0.3247895) [4]: 6 45 66 74 
#> 90: (0.6620601,0.9973517) [5]: 12 26 38 60 97 
#> 91: (0.5608448,0.1008173) [5]: 3 43 46 56 98 
#> 92: (0.8788632,0.7704715) [7]: 5 14 32 36 57 64 88 
#> 93: (0.4210088,0.8117368) [8]: 19 33 35 44 58 60 63 75 
#> 94: (0.5484703,0.4232588) [6]: 52 68 78 79 82 95 
#> 95: (0.5141424,0.5761063) [7]: 41 53 62 63 79 82 94 
#> 96: (0.9148878,0.4711823) [3]: 39 73 85 
#> 97: (0.9278724,0.9120003) [6]: 8 36 38 51 57 90 
#> 98: (0.6175188,0.1364148) [5]: 40 43 56 69 91 
#> 99: (0.6236884,0.6068507) [4]: 41 53 64 76 
#> 100: (0.1475072,0.1576683) [8]: 4 18 25 29 37 48 66 86 
#> number of nodes: 100 
#> number of arcs: 282 
#> number of boundary nodes: 15 
#> boundary nodes:  9 12 17 22 31 37 50 51 54 59 61 83 88 90 97 
#> number of triangles: 183 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.13089193 0.12809283 0.04629755 0.05927735 0.03288818 0.01629641 0.13596037 
#>                                                                              
#> 0.04606409 0.14044284 0.15687235 0.10192631 0.19544339 0.29878415 0.15684119 
#>                                                                              
#> 0.04395441 0.08118555 0.17223005 0.11428780 0.19335851 0.13680088 0.12004879 
#>                                                                              
#> 0.17948324 0.15496606 0.08452440 0.10679648 0.02370085 0.11697792 0.05170679 
#>                                                                              
#> 0.20286760 0.11375381 0.04066383 0.16848099 0.16282290 0.17532263 0.04649631 
#>                                                                              
#> 0.02030461 0.01840848 0.06926380 0.51624076 0.11799092 0.05080386 0.10396012 
#>                                                                              
#> 0.06775705 0.05062653 0.19784706 0.12424224 0.18914317 0.14354965 0.12155804 
#>                                                                              
#> 0.23908586 0.20196449 0.03144819 0.30281654 0.09652397 0.18746082 0.13588268 
#>                                                                              
#> 0.11885734 0.14449768 0.09439074 0.02193903 0.18238659 0.10787219 0.23759029 
#>                                                                              
#> 0.13456799 0.05451477 0.16603002 0.08150230 0.07467894 0.06138997 0.19846903 
#>                                                                              
#> 0.19593714 0.11801698 0.07693483 0.08274478 0.20237482 0.12857190 0.11153839 
#>                                                                              
#> 0.14256603 0.05577809 0.03410240 0.11069658 0.04259406 0.16088352 0.20665045 
#>                                                                              
#> 0.02082428 0.19799295 0.57807260 0.10672824 0.10209147 0.22058057 0.04602381 
#>                                                                              
#> 0.07354343 0.17232532 0.22051990 0.24818808 0.02771312 0.07775453 0.09515508 
#>                                                                              
#> 0.03809756 0.10477212 0.05474586 0.12986020 0.12095063 0.11703251 0.03188046 
#>                                                                              
#> 0.20102594 0.00969843 0.01837164 0.19287419 0.18620605 0.02610865 0.11914566 
#>                                                                              
#> 0.06448543 0.14080662 0.27917924 0.07380911 0.01760022 0.16490612 0.12671236 
#>                                                                              
#> 0.05887306 0.02920493 0.08530457 0.18115984 0.17932943 0.07552973 0.06692628 
#>                                                                              
#> 0.04358628 0.02673442 0.20527350 0.08346187 0.10787924 0.22517138 0.22898404 
#>                                                                              
#> 0.16841336 0.37919232 0.32121450 0.06480019 0.13795005 0.09071033 0.15665493 
#>                                                                              
#> 0.12653417 0.06248137 0.02014771 0.05285416 0.03586755 0.02304343 0.06968141 
#>                                                                              
#> 0.04800915 0.11509637 0.14274090 0.19806685 0.18047737 0.20053089 0.06971376 
#>                                                                              
#> 0.04731814 0.10435438 0.04527066 0.16091109 0.12984447 0.08377643 0.05795347 
#>                                                                              
#> 0.17017505 0.10486528 0.07801318 0.03880405 0.05647746 0.09002930 0.05189472 
#>                                                                              
#> 0.18901096 0.12443860 0.09621613 0.12076143 0.10200384 0.21420510 0.02147248 
#>                                                                              
#> 0.01834076 0.01123176 0.05803437 0.17489200 0.07743734 0.13747632 0.16692259 
#>                                                                              
#> 0.18092167 0.19776954 0.20040061 0.07658536 0.08262312 0.14327549 0.17591927 
#>                                                                              
#> 0.24647022 0.23211851 0.19401319 0.20774933 0.14179871 0.15009964 0.19893214 
#>                                                                              
#> 0.13352343 0.15738954 0.13925768 0.07602081 0.05767327 0.18542330 0.15808894 
#>                                                                              
#> 0.10654638 0.17502222 0.12138670 0.07966970 0.17354309 0.15785285 0.15825017 
#>                                                                              
#> 0.15486144 0.07445878 0.10560397 0.09603830 0.10639014 0.15326882 0.17929752 
#>                                                                              
#> 0.18029867 0.13234312 0.05463012 0.03763806 0.05674848 0.15588392 0.14001972 
#>                                                                              
#> 0.48640953 0.29964535 0.13016530 0.13924662 0.15148251 0.04510319 0.08867127 
#>                                                                              
#> 0.07819609 0.10035842 0.16887668 0.22099873 0.10657166 0.08469736 0.10013472 
#>                                                                              
#> 0.15788105 0.09393423 0.13133866 0.11143764 0.11230700 0.12723714 0.18597830 
#>                                                                              
#> 0.18202150 0.18987973 0.17202460 0.03542798 0.12642670 0.13032796 0.10524238 
#>                                                                              
#> 0.17366808 0.17005085 0.03228991 0.04780895 0.07543735 0.03946892 0.06027351 
#>                                                                              
#> 0.13308109 0.13257636 0.02931618 0.18350576 0.08911910 0.01970579 0.10641934 
#>                                                                              
#> 0.20472674 0.17543498 0.16384786 0.12013712 0.24420829 0.07553145 0.48578450 
#>                                                                              
#> 0.16822431 0.09751506 0.02328550 0.08468530 0.69079183 0.22039670 0.13052540 
#>                       
#> 0.24426676 0.23597062 
#> 
#> $alpha
#> [1] 0.3270898
#> 
#> attr(,"class")
#> [1] "scree"