Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2           mx1
#>    50   17 0.421788754 0.077305250 0.588845603 0.005275755   0.509446587
#>    17   76 0.588845603 0.005275755 0.625330802 0.029545397   0.549929505
#>    76   20 0.625330802 0.029545397 0.654153003 0.197398296   0.561273185
#>    20   19 0.654153003 0.197398296 0.633377521 0.231518111   0.549545682
#>    19   28 0.633377521 0.231518111 0.608515825 0.269428294   0.529180107
#>    28   46 0.608515825 0.269428294 0.410598509 0.191206712   0.522610528
#>    46    8 0.410598509 0.191206712 0.407467630 0.149861219   0.456702715
#>     8    7 0.407467630 0.149861219 0.394541306 0.140030206   0.417561601
#>     7   50 0.394541306 0.140030206 0.421788754 0.077305250   0.411365707
#>    75   29 0.808102576 0.817776937 0.817309464 0.878523713   0.746055472
#>    29   66 0.817309464 0.878523713 0.853927250 0.953367776   0.724357169
#>    66   70 0.853927250 0.953367776 0.557693365 0.888830151   0.676787720
#>    70    6 0.557693365 0.888830151 0.628460612 0.763646458   0.602255013
#>     6   75 0.628460612 0.763646458 0.808102576 0.817776937   0.724671783
#>    77   62 0.990190287 0.952519390 0.982533878 0.426001094   4.814463193
#>    62   56 0.982533878 0.426001094 0.862137303 0.399992607   0.943350662
#>    56   85 0.862137303 0.399992607 0.816273847 0.178898223   0.892737920
#>    85   22 0.816273847 0.178898223 0.884669450 0.143890305   0.897449731
#>    22   89 0.884669450 0.143890305 0.977838344 0.036102643   0.973617637
#>    60   77 0.939447983 0.553345795 0.990190287 0.952519390   1.053059544
#>    77   51 0.990190287 0.952519390 0.931043665 0.965007655   0.936906076
#>    51  100 0.931043665 0.965007655 0.817272621 0.829544889   0.931076612
#>   100   75 0.817272621 0.829544889 0.808102576 0.817776937   0.832431816
#>    75   90 0.808102576 0.817776937 0.761565677 0.674961341   0.764713986
#>    90   49 0.761565677 0.674961341 0.800728639 0.658465494   0.805446004
#>    49   60 0.800728639 0.658465494 0.939447983 0.553345795   0.931937546
#>    68   27 0.391000472 0.606623440 0.409067002 0.662215082   0.377269237
#>    27   92 0.409067002 0.662215082 0.319517027 0.710950898   0.346665629
#>    92   42 0.319517027 0.710950898 0.227365315 0.657998734   0.305631018
#>    42   21 0.227365315 0.657998734 0.220637632 0.626503431   0.304496205
#>    21   91 0.220637632 0.626503431 0.197086987 0.554282839   0.296297226
#>    91   94 0.197086987 0.554282839 0.356234303 0.477675374   0.297178336
#>    94   48 0.356234303 0.477675374 0.413333737 0.537396483   0.342406279
#>    48   68 0.413333737 0.537396483 0.391000472 0.606623440   0.399294908
#>    70   93 0.557693365 0.888830151 0.570439810 0.703127153   0.536160917
#>    93   43 0.570439810 0.703127153 0.609317403 0.728188553   0.544474639
#>    43   81 0.609317403 0.728188553 0.670797237 0.679766573   0.656918651
#>    81   90 0.670797237 0.679766573 0.761565677 0.674961341   0.720420485
#>    84   61 0.338092592 0.272181274 0.267086300 0.085120146   0.302299835
#>    61   55 0.267086300 0.085120146 0.343033909 0.015417869   0.340854069
#>    55   50 0.343033909 0.015417869 0.421788754 0.077305250   0.352463779
#>    46   84 0.410598509 0.191206712 0.338092592 0.272181274   0.327024262
#>    69   82 0.142754792 0.464170945 0.057379641 0.481532350   0.099578057
#>    82   23 0.057379641 0.481532350 0.022783088 0.325699131   0.020599182
#>    23   88 0.022783088 0.325699131 0.120452868 0.397359418   0.059610391
#>    88   69 0.120452868 0.397359418 0.142754792 0.464170945   0.136128885
#>    93   54 0.570439810 0.703127153 0.594771147 0.539060839   0.574751586
#>    54   81 0.594771147 0.539060839 0.670797237 0.679766573   0.637155055
#>    53   78 0.000000000 0.037007413 0.011759334 0.102478752  -0.320423293
#>    78   15 0.011759334 0.102478752 0.081474201 0.148958476   0.026112288
#>    15   71 0.081474201 0.148958476 0.117050147 0.275202380   0.073179945
#>    71   23 0.117050147 0.275202380 0.022783088 0.325699131   0.048617074
#>    23   96 0.022783088 0.325699131 0.027484852 0.775036926  -1.821157119
#>    96   26 0.027484852 0.775036926 0.009884877 0.954842778  -2.640255351
#>    26   53 0.009884877 0.954842778 0.000000000 0.037007413 -45.330654154
#>    83   56 0.896427428 0.537330028 0.862137303 0.399992607   0.879389740
#>    56   40 0.862137303 0.399992607 0.974112745 0.429031214   0.909338157
#>    40   62 0.974112745 0.429031214 0.982533878 0.426001094   1.001063084
#>    62   60 0.982533878 0.426001094 0.939447983 0.553345795   1.061716220
#>    60   83 0.939447983 0.553345795 0.896427428 0.537330028   0.922606247
#>    56   98 0.862137303 0.399992607 0.730891059 0.328598232   0.832989805
#>    98   20 0.730891059 0.328598232 0.654153003 0.197398296   0.723619175
#>    20   74 0.654153003 0.197398296 0.731047676 0.129751888   0.723818653
#>    74   85 0.731047676 0.129751888 0.816273847 0.178898223   0.774718286
#>    93   63 0.570439810 0.703127153 0.479093896 0.545491166   0.516865704
#>    63   24 0.479093896 0.545491166 0.463052611 0.386324318   0.506858300
#>    24   31 0.463052611 0.386324318 0.622992310 0.500472855   0.543071484
#>    31   54 0.622992310 0.500472855 0.594771147 0.539060839   0.567030190
#>    78   79 0.011759334 0.102478752 0.035091991 0.026561249   0.076535862
#>    79   87 0.035091991 0.026561249 0.119850067 0.150919346   0.104738780
#>    87   67 0.119850067 0.150919346 0.128958828 0.263506224   0.097624938
#>    67   71 0.128958828 0.263506224 0.117050147 0.275202380   0.073800004
#>    42   36 0.227365315 0.657998734 0.215910105 0.847880325   0.161229420
#>    36   95 0.215910105 0.847880325 0.080860026 0.856704108   0.142562212
#>    95   96 0.080860026 0.856704108 0.027484852 0.775036926   0.134911286
#>    96   34 0.027484852 0.775036926 0.044107130 0.623726352   0.005241968
#>    34   91 0.044107130 0.623726352 0.197086987 0.554282839   0.123194672
#>    50   25 0.421788754 0.077305250 0.366540400 0.009435556   0.477103559
#>    25   18 0.366540400 0.009435556 0.591025612 0.000000000   0.476729401
#>    18   76 0.591025612 0.000000000 0.625330802 0.029545397   0.611098047
#>    97   76 0.707941971 0.006367211 0.625330802 0.029545397   0.651598235
#>    98   28 0.730891059 0.328598232 0.608515825 0.269428294   0.675784605
#>    76   74 0.625330802 0.029545397 0.731047676 0.129751888   0.647370705
#>    97   89 0.707941971 0.006367211 0.977838344 0.036102643   0.843656462
#>    74   97 0.731047676 0.129751888 0.707941971 0.006367211   0.815847578
#>    82   34 0.057379641 0.481532350 0.044107130 0.623726352  -0.429295336
#>    71   88 0.117050147 0.275202380 0.120452868 0.397359418   0.089545814
#>    63   48 0.479093896 0.545491166 0.413333737 0.537396483   0.455316381
#>    94   39 0.356234303 0.477675374 0.411222588 0.334417334   0.389265678
#>    39   58 0.411222588 0.334417334 0.473859348 0.310625804   0.451478356
#>    58   47 0.473859348 0.310625804 0.652073691 0.406324342   0.561224714
#>    47   31 0.652073691 0.406324342 0.622992310 0.500472855   0.554246272
#>    55   79 0.343033909 0.015417869 0.035091991 0.026561249   0.161231566
#>    96   38 0.027484852 0.775036926 0.051667859 0.866037408  -0.088157115
#>    38   52 0.051667859 0.866037408 0.142991089 0.877598027   0.088973970
#>    52   65 0.142991089 0.877598027 0.217682858 1.000000000   0.113285435
#>    68   63 0.391000472 0.606623440 0.479093896 0.545491166   0.461651166
#>    93   44 0.570439810 0.703127153 0.405964106 0.719022074   0.486642772
#>    44   92 0.405964106 0.719022074 0.319517027 0.710950898   0.365231329
#>    98   47 0.730891059 0.328598232 0.652073691 0.406324342   0.654661352
#>    58   46 0.473859348 0.310625804 0.410598509 0.191206712   0.517089164
#>    51   37 0.931043665 0.965007655 0.864661331 0.949070318   0.921153038
#>    37   66 0.864661331 0.949070318 0.853927250 0.953367776   0.843607512
#>    44   35 0.405964106 0.719022074 0.472598314 0.921277142   0.402856907
#>    35   80 0.472598314 0.921277142 0.276691518 0.832191656   0.375486676
#>    80   36 0.276691518 0.832191656 0.215910105 0.847880325   0.232652964
#>    47   86 0.652073691 0.406324342 0.695281570 0.478803304   0.649369729
#>    86   54 0.695281570 0.478803304 0.594771147 0.539060839   0.684815260
#>    66   41 0.853927250 0.953367776 0.401843244 0.988615034   0.659659858
#>    41   35 0.401843244 0.988615034 0.472598314 0.921277142   0.428691856
#>    35   99 0.472598314 0.921277142 0.482677402 0.918400379   0.481382509
#>    99   70 0.482677402 0.918400379 0.557693365 0.888830151   0.534363308
#>    64   45 0.287348654 0.376388089 0.158316665 0.387842903   0.220745622
#>    45   72 0.158316665 0.387842903 0.219276071 0.284729119   0.217928962
#>    72   84 0.219276071 0.284729119 0.338092592 0.272181274   0.276550409
#>    84   64 0.338092592 0.272181274 0.287348654 0.376388089   0.343664646
#>    82   91 0.057379641 0.481532350 0.197086987 0.554282839   0.106445057
#>    41   80 0.401843244 0.988615034 0.276691518 0.832191656   0.370810780
#>    44   99 0.405964106 0.719022074 0.482677402 0.918400379   0.448333470
#>    80   57 0.276691518 0.832191656 0.307892188 0.979832164   0.264828329
#>    57   65 0.307892188 0.979832164 0.217682858 1.000000000   0.247956296
#>    52   95 0.142991089 0.877598027 0.080860026 0.856704108   0.143266659
#>    51   66 0.931043665 0.965007655 0.853927250 0.953367776   0.884271529
#>    94   64 0.356234303 0.477675374 0.287348654 0.376388089   0.363860778
#>    51   65 0.931043665 0.965007655 0.217682858 1.000000000   1.593711862
#>    71   72 0.117050147 0.275202380 0.219276071 0.284729119   0.164265944
#>    64   69 0.287348654 0.376388089 0.142754792 0.464170945   0.228133609
#>    98   86 0.730891059 0.328598232 0.695281570 0.478803304   0.731532685
#>    90   86 0.761565677 0.674961341 0.695281570 0.478803304   0.741454637
#>    86   83 0.695281570 0.478803304 0.896427428 0.537330028   0.784456200
#>    86   81 0.695281570 0.478803304 0.670797237 0.679766573   0.687532135
#>    61   59 0.267086300 0.085120146 0.226560606 0.045952323   0.283881811
#>    59   79 0.226560606 0.045952323 0.035091991 0.026561249   0.174139299
#>    83   73 0.896427428 0.537330028 0.818037812 0.432504516   0.878814465
#>    73   98 0.818037812 0.432504516 0.730891059 0.328598232   0.798752312
#>    61   87 0.267086300 0.085120146 0.119850067 0.150919346   0.194299848
#>    72   87 0.219276071 0.284729119 0.119850067 0.150919346   0.220992603
#>    91   69 0.197086987 0.554282839 0.142754792 0.464170945   0.242488562
#>     1   17 0.445201599 0.116101332 0.588845603 0.005275755   0.509446587
#>     1   76 0.445201599 0.116101332 0.625330802 0.029545397   0.549929505
#>     1   20 0.445201599 0.116101332 0.654153003 0.197398296   0.561273185
#>     1   19 0.445201599 0.116101332 0.633377521 0.231518111   0.549545682
#>     1   28 0.445201599 0.116101332 0.608515825 0.269428294   0.529180107
#>     1   46 0.445201599 0.116101332 0.410598509 0.191206712   0.522610528
#>     1    8 0.445201599 0.116101332 0.407467630 0.149861219   0.456702715
#>     1    7 0.445201599 0.116101332 0.394541306 0.140030206   0.417561601
#>     1   50 0.445201599 0.116101332 0.421788754 0.077305250   0.411365707
#>     2   29 0.672382896 0.850478630 0.817309464 0.878523713   0.746055472
#>     2   66 0.672382896 0.850478630 0.853927250 0.953367776   0.724357169
#>     2   70 0.672382896 0.850478630 0.557693365 0.888830151   0.676787720
#>     2    6 0.672382896 0.850478630 0.628460612 0.763646458   0.602255013
#>     2   75 0.672382896 0.850478630 0.808102576 0.817776937   0.724671783
#>     3   62 1.000000000 0.213299094 0.982533878 0.426001094   4.814463193
#>     3   56 1.000000000 0.213299094 0.862137303 0.399992607   0.943350662
#>     3   85 1.000000000 0.213299094 0.816273847 0.178898223   0.892737920
#>     3   22 1.000000000 0.213299094 0.884669450 0.143890305   0.897449731
#>     3   89 1.000000000 0.213299094 0.977838344 0.036102643   0.973617637
#>     4   77 0.836902255 0.782542103 0.990190287 0.952519390   1.053059544
#>     4   51 0.836902255 0.782542103 0.931043665 0.965007655   0.936906076
#>     4  100 0.836902255 0.782542103 0.817272621 0.829544889   0.931076612
#>     4   75 0.836902255 0.782542103 0.808102576 0.817776937   0.832431816
#>     4   90 0.836902255 0.782542103 0.761565677 0.674961341   0.764713986
#>     4   49 0.836902255 0.782542103 0.800728639 0.658465494   0.805446004
#>     4   60 0.836902255 0.782542103 0.939447983 0.553345795   0.931937546
#>     5   27 0.385945992 0.605049540 0.409067002 0.662215082   0.377269237
#>     5   92 0.385945992 0.605049540 0.319517027 0.710950898   0.346665629
#>     5   42 0.385945992 0.605049540 0.227365315 0.657998734   0.305631018
#>     5   21 0.385945992 0.605049540 0.220637632 0.626503431   0.304496205
#>     5   91 0.385945992 0.605049540 0.197086987 0.554282839   0.296297226
#>     5   94 0.385945992 0.605049540 0.356234303 0.477675374   0.297178336
#>     5   48 0.385945992 0.605049540 0.413333737 0.537396483   0.342406279
#>     5   68 0.385945992 0.605049540 0.391000472 0.606623440   0.399294908
#>     6   93 0.628460612 0.763646458 0.570439810 0.703127153   0.536160917
#>     6   43 0.628460612 0.763646458 0.609317403 0.728188553   0.544474639
#>     6   81 0.628460612 0.763646458 0.670797237 0.679766573   0.656918651
#>     6   90 0.628460612 0.763646458 0.761565677 0.674961341   0.720420485
#>     7   61 0.394541306 0.140030206 0.267086300 0.085120146   0.302299835
#>     7   55 0.394541306 0.140030206 0.343033909 0.015417869   0.340854069
#>     7   46 0.394541306 0.140030206 0.410598509 0.191206712   0.379863600
#>     7   84 0.394541306 0.140030206 0.338092592 0.272181274   0.327024262
#>     9   82 0.100672358 0.426829551 0.057379641 0.481532350   0.099578057
#>     9   23 0.100672358 0.426829551 0.022783088 0.325699131   0.020599182
#>     9   88 0.100672358 0.426829551 0.120452868 0.397359418   0.059610391
#>     9   69 0.100672358 0.426829551 0.142754792 0.464170945   0.136128885
#>    10   93 0.625147749 0.686267091 0.570439810 0.703127153   0.599603988
#>    10   54 0.625147749 0.686267091 0.594771147 0.539060839   0.574751586
#>    10   81 0.625147749 0.686267091 0.670797237 0.679766573   0.637155055
#>    10   43 0.625147749 0.686267091 0.609317403 0.728188553   0.653363004
#>    11   78 0.003240751 0.207446833 0.011759334 0.102478752  -0.320423293
#>    11   15 0.003240751 0.207446833 0.081474201 0.148958476   0.026112288
#>    11   71 0.003240751 0.207446833 0.117050147 0.275202380   0.073179945
#>    11   23 0.003240751 0.207446833 0.022783088 0.325699131   0.048617074
#>    11   96 0.003240751 0.207446833 0.027484852 0.775036926  -1.821157119
#>    11   26 0.003240751 0.207446833 0.009884877 0.954842778  -2.640255351
#>    11   53 0.003240751 0.207446833 0.000000000 0.037007413 -45.330654154
#>    12   56 0.936645428 0.510241520 0.862137303 0.399992607   0.879389740
#>    12   40 0.936645428 0.510241520 0.974112745 0.429031214   0.909338157
#>    12   62 0.936645428 0.510241520 0.982533878 0.426001094   1.001063084
#>    12   60 0.936645428 0.510241520 0.939447983 0.553345795   1.061716220
#>    12   83 0.936645428 0.510241520 0.896427428 0.537330028   0.922606247
#>    13   98 0.793496157 0.198005879 0.730891059 0.328598232   0.832989805
#>    13   20 0.793496157 0.198005879 0.654153003 0.197398296   0.723619175
#>    13   74 0.793496157 0.198005879 0.731047676 0.129751888   0.723818653
#>    13   85 0.793496157 0.198005879 0.816273847 0.178898223   0.774718286
#>    13   56 0.793496157 0.198005879 0.862137303 0.399992607   0.882135455
#>    14   63 0.544382087 0.541569086 0.479093896 0.545491166   0.516865704
#>    14   24 0.544382087 0.541569086 0.463052611 0.386324318   0.506858300
#>    14   31 0.544382087 0.541569086 0.622992310 0.500472855   0.543071484
#>    14   54 0.544382087 0.541569086 0.594771147 0.539060839   0.567030190
#>    14   93 0.544382087 0.541569086 0.570439810 0.703127153   0.573530613
#>    15   79 0.081474201 0.148958476 0.035091991 0.026561249   0.076535862
#>    15   87 0.081474201 0.148958476 0.119850067 0.150919346   0.104738780
#>    15   67 0.081474201 0.148958476 0.128958828 0.263506224   0.097624938
#>    16   36 0.082581794 0.668526689 0.215910105 0.847880325   0.161229420
#>    16   95 0.082581794 0.668526689 0.080860026 0.856704108   0.142562212
#>    16   96 0.082581794 0.668526689 0.027484852 0.775036926   0.134911286
#>    16   34 0.082581794 0.668526689 0.044107130 0.623726352   0.005241968
#>    16   91 0.082581794 0.668526689 0.197086987 0.554282839   0.123194672
#>    16   21 0.082581794 0.668526689 0.220637632 0.626503431   0.140962009
#>    16   42 0.082581794 0.668526689 0.227365315 0.657998734   0.154524844
#>    17   25 0.588845603 0.005275755 0.366540400 0.009435556   0.477103559
#>    17   18 0.588845603 0.005275755 0.591025612 0.000000000   0.476729401
#>    19   98 0.633377521 0.231518111 0.730891059 0.328598232   0.708313870
#>    22   74 0.884669450 0.143890305 0.731047676 0.129751888   0.812436106
#>    22   97 0.884669450 0.143890305 0.707941971 0.006367211   0.815847578
#>    23   34 0.022783088 0.325699131 0.044107130 0.623726352  -0.429295336
#>    24   48 0.463052611 0.386324318 0.413333737 0.537396483   0.455316381
#>    24   94 0.463052611 0.386324318 0.356234303 0.477675374   0.434000050
#>    24   39 0.463052611 0.386324318 0.411222588 0.334417334   0.389265678
#>    24   58 0.463052611 0.386324318 0.473859348 0.310625804   0.451478356
#>    24   47 0.463052611 0.386324318 0.652073691 0.406324342   0.561224714
#>    25   55 0.366540400 0.009435556 0.343033909 0.015417869   0.368068504
#>    25   79 0.366540400 0.009435556 0.035091991 0.026561249   0.161231566
#>    26   38 0.009884877 0.954842778 0.051667859 0.866037408  -0.088157115
#>    26   52 0.009884877 0.954842778 0.142991089 0.877598027   0.088973970
#>    26   65 0.009884877 0.954842778 0.217682858 1.000000000   0.113285435
#>    27   63 0.409067002 0.662215082 0.479093896 0.545491166   0.461651166
#>    27   93 0.409067002 0.662215082 0.570439810 0.703127153   0.501068007
#>    27   44 0.409067002 0.662215082 0.405964106 0.719022074   0.486642772
#>    28   47 0.608515825 0.269428294 0.652073691 0.406324342   0.654661352
#>    28   58 0.608515825 0.269428294 0.473859348 0.310625804   0.562432940
#>    29   37 0.817309464 0.878523713 0.864661331 0.949070318   0.921153038
#>    29  100 0.817309464 0.878523713 0.817272621 0.829544889   0.773667209
#>    29   51 0.817309464 0.878523713 0.931043665 0.965007655   0.925741814
#>    30   44 0.292731381 0.806074716 0.405964106 0.719022074   0.357330470
#>    30   35 0.292731381 0.806074716 0.472598314 0.921277142   0.402856907
#>    30   80 0.292731381 0.806074716 0.276691518 0.832191656   0.375486676
#>    30   36 0.292731381 0.806074716 0.215910105 0.847880325   0.232652964
#>    30   42 0.292731381 0.806074716 0.227365315 0.657998734   0.213771531
#>    30   92 0.292731381 0.806074716 0.319517027 0.710950898   0.241374378
#>    31   86 0.622992310 0.500472855 0.695281570 0.478803304   0.649369729
#>    32   41 0.479053095 0.947387478 0.401843244 0.988615034   0.659659858
#>    32   35 0.479053095 0.947387478 0.472598314 0.921277142   0.428691856
#>    32   99 0.479053095 0.947387478 0.482677402 0.918400379   0.481382509
#>    32   70 0.479053095 0.947387478 0.557693365 0.888830151   0.534363308
#>    32   66 0.479053095 0.947387478 0.853927250 0.953367776   0.663887451
#>    33   45 0.278721766 0.321311078 0.158316665 0.387842903   0.220745622
#>    33   72 0.278721766 0.321311078 0.219276071 0.284729119   0.217928962
#>    33   84 0.278721766 0.321311078 0.338092592 0.272181274   0.276550409
#>    33   64 0.278721766 0.321311078 0.287348654 0.376388089   0.343664646
#>    36   57 0.215910105 0.847880325 0.307892188 0.979832164   0.264828329
#>    36   65 0.215910105 0.847880325 0.217682858 1.000000000   0.247956296
#>    36   52 0.215910105 0.847880325 0.142991089 0.877598027   0.204451319
#>    38   95 0.051667859 0.866037408 0.080860026 0.856704108   0.052141102
#>    39   64 0.411222588 0.334417334 0.287348654 0.376388089   0.363860778
#>    39   84 0.411222588 0.334417334 0.338092592 0.272181274   0.343883852
#>    39   46 0.411222588 0.334417334 0.410598509 0.191206712   0.409106925
#>    41   57 0.401843244 0.988615034 0.307892188 0.979832164   0.327342303
#>    41   51 0.401843244 0.988615034 0.931043665 0.965007655   0.717402455
#>    41   65 0.401843244 0.988615034 0.217682858 1.000000000   1.593711862
#>    44   70 0.405964106 0.719022074 0.557693365 0.888830151   0.495955278
#>    45   71 0.158316665 0.387842903 0.117050147 0.275202380   0.125184451
#>    45   69 0.158316665 0.387842903 0.142754792 0.464170945   0.228133609
#>    45   88 0.158316665 0.387842903 0.120452868 0.397359418   0.147632026
#>    49   86 0.800728639 0.658465494 0.695281570 0.478803304   0.741454637
#>    49   83 0.800728639 0.658465494 0.896427428 0.537330028   0.784456200
#>    55   59 0.343033909 0.015417869 0.226560606 0.045952323   0.283881811
#>    56   73 0.862137303 0.399992607 0.818037812 0.432504516   0.878814465
#>    59   87 0.226560606 0.045952323 0.119850067 0.150919346   0.194299848
#>    61   72 0.267086300 0.085120146 0.219276071 0.284729119   0.269471878
#>    64   91 0.287348654 0.376388089 0.197086987 0.554282839   0.255535362
#>    67   72 0.128958828 0.263506224 0.219276071 0.284729119   0.191106953
#>    73   86 0.818037812 0.432504516 0.695281570 0.478803304   0.787394209
#>     3   77 1.000000000 0.213299094 0.990190287 0.952519390   4.814463193
#>    18   97 0.591025612 0.000000000 0.707941971 0.006367211   0.651598235
#>    53   79 0.000000000 0.037007413 0.035091991 0.026561249   0.027679954
#>          my1           mx2          my2 bp1 bp2
#>   0.05086777   0.477103559  -0.02414489   0   0
#>   0.10333877   0.611098047   0.01138247   0   0
#>   0.12694579   0.647370705   0.11216190   0   0
#>   0.15708810   0.708313870   0.25376171   0   0
#>   0.19029221   0.675784605   0.28643621   0   0
#>   0.19728971   0.517089164   0.21125994   0   0
#>   0.16692419   0.379863600   0.17274282   0   0
#>   0.12317554   0.379863600   0.17274282   0   0
#>   0.11005809   0.352463779   0.08447135   0   0
#>   0.85825200   0.773667209   0.85406712   0   0
#>   0.97038077   0.843607512   0.91203702   0   0
#>   1.05431542   0.663887451   1.11352889   0   0
#>   0.83142671   0.536160917   0.79406324   0   0
#>   0.76950468   0.725851755   0.76558873   0   0
#>   0.63359360   1.527626726   0.68138940   0   0
#>   0.31571543   0.941378978   0.32484260   0   0
#>   0.27834075   0.882135455   0.28054011   0   0
#>   0.25317619   0.812436106   0.08708354   0   0
#>   0.12661461   0.843656462   0.01427949   0   0
#>   0.74171561   1.527626726   0.68138940   0   0
#>   0.84646453   4.354341704  17.03204090   0   1
#>   0.84947219   0.925741814   0.85395272   0   0
#>   0.80827545   0.773667209   0.85406712   0   0
#>   0.75292535   0.725851755   0.76558873   0   0
#>   0.72440156   0.741454637   0.57247899   0   0
#>   0.68752389   0.887054420   0.62829467   0   0
#>   0.64181742   0.461651166   0.61439443   0   0
#>   0.65419526   0.365231329   0.68830894   0   0
#>   0.62845540   0.241374378   0.74028009   0   0
#>   0.62505668   0.154524844   0.65709192   0   0
#>   0.56188120   0.140962009   0.61253496   0   0
#>   0.55860335   0.255535362   0.47209263   0   0
#>   0.54805334   0.434000050   0.46048041   0   0
#>   0.57108336   0.440912671   0.58450964   0   0
#>   0.79406324   0.495955278   0.79130357   0   0
#>   0.78609274   0.599603988   0.70057093   0   0
#>   0.72538585   0.653363004   0.72087136   0   0
#>   0.75743708   0.711170649   0.58271230   0   0
#>   0.17876064   0.269471878   0.19122176   0   0
#>   0.08927012   0.283881811   0.02719314   0   0
#>   0.08447135   0.368068504   0.06461354   0   0
#>   0.18932175   0.409106925   0.26281988   0   0
#>   0.47044620   0.114285745   0.54277163   0   0
#>   0.40794098  -0.429295336   0.50782212   0   0
#>   0.37789508   0.089545814   0.33709443   0   0
#>   0.42925470   0.147632026   0.42541491   0   0
#>   0.61992925   0.573530613   0.61974818   0   0
#>   0.60705204   0.687532135   0.57983231   0   0
#>   0.12835049   0.027679954   0.06582752   0   0
#>   0.15647325   0.076535862   0.08084301   0   0
#>   0.21943049   0.073800004   0.21925575   0   0
#>   0.26068889   0.089545814   0.33709443   0   0
#>   0.56968718  -1.203617743   0.56322540   0   0
#>   0.60467423  -0.088157115   0.85448182   0   0
#>   0.98417909 -46.420047812   0.99591161   0   1
#>   0.46863451   0.878814465   0.46877814   0   0
#>   0.44839485   0.941378978   0.32484260   0   0
#>   0.49071321   0.941378978   0.32484260   0   0
#>   0.52375291   1.527626726   0.68138940   0   0
#>   0.53279757   0.887054420   0.62829467   0   0
#>   0.29724128   0.798752312   0.36018100   0   0
#>   0.24480973   0.708313870   0.25376171   0   0
#>   0.19906141   0.647370705   0.11216190   0   0
#>   0.15249117   0.812436106   0.08708354   0   0
#>   0.62888767   0.501068007   0.63804202   0   0
#>   0.46230122   0.455316381   0.46749576   0   0
#>   0.44332990   0.554246272   0.42767230   0   0
#>   0.48915890   0.684815260   0.57530052   0   0
#>   0.08084301   0.027679954   0.06582752   0   0
#>   0.07015556   0.129072863   0.05357031   0   0
#>   0.20937936   0.191106953   0.20181627   0   0
#>   0.21925575   0.165152729   0.31226840   0   0
#>   0.74929521   0.213771531   0.75246498   0   0
#>   0.76317208   0.143266659   0.77395381   0   0
#>   0.76310208   0.052141102   0.81719814   0   0
#>   0.69602512  -1.203617743   0.56322540   0   0
#>   0.59472698   0.106445057   0.55782859   0   0
#>  -0.02414489   0.368068504   0.06461354   0   0
#>  -0.04414038   0.289822454  -4.49092072   0   1
#>   0.01138247   0.651598235  -0.03564234   0   0
#>  -0.03564234   0.682630714   0.07496293   0   0
#>   0.28643621   0.654661352   0.33012331   0   0
#>   0.11216190   0.682630714   0.07496293   0   0
#>   0.01427949   1.246967712  -3.64641204   0   1
#>   0.05001599   0.682630714   0.07496293   0   0
#>   0.50782212   0.106445057   0.55782859   0   0
#>   0.33709443   0.125184451   0.33610171   0   0
#>   0.46749576   0.440912671   0.58450964   0   0
#>   0.40817177   0.363860778   0.39842033   0   0
#>   0.34605133   0.419846501   0.26277308   0   0
#>   0.36171875   0.562432940   0.35946873   0   0
#>   0.42767230   0.649369729   0.45705483   0   0
#>  -0.74811734   0.174139299  -0.39141835   0   0
#>   0.85448182   0.052141102   0.81719814   0   0
#>   0.93782210   0.089465498   0.93393927   0   0
#>   0.97971501   0.204451319   0.92408403   0   0
#>   0.61439443   0.440912671   0.58450964   0   0
#>   0.69494064   0.495955278   0.79130357   0   0
#>   0.68830894   0.357330470   0.77293183   0   0
#>   0.33012331   0.731532685   0.40807389   0   0
#>   0.21125994   0.419846501   0.26277308   0   0
#>   0.85998738   0.884271529   1.01360655   0   0
#>   0.91203702   0.884271529   1.01360655   0   0
#>   0.83214982   0.448333470   0.81716728   0   0
#>   0.87488329   0.370810780   0.88516600   0   0
#>   0.78716118   0.264828329   0.91181574   0   0
#>   0.45705483   0.731532685   0.40807389   0   0
#>   0.57530052   0.687532135   0.57983231   0   0
#>   1.37853497   0.717402455   2.11914590   0   0
#>   0.94598435   0.370810780   0.88516600   0   0
#>   0.93295860   0.448333470   0.81716728   0   0
#>   0.93958288   0.481134024   0.80454686   0   0
#>   0.35860620   0.228133609   0.44182769   0   0
#>   0.35350879   0.164265944   0.32178397   0   0
#>   0.25824892   0.269471878   0.19122176   0   0
#>   0.33935300   0.343883852   0.33945974   0   0
#>   0.55782859   0.114285745   0.54277163   0   0
#>   0.88516600   0.363587718   0.89094505   0   0
#>   0.81716728   0.481134024   0.80454686   0   0
#>   0.91181574   0.363587718   0.89094505   0   0
#>   0.92357704   0.327342303   1.27866512   0   0
#>   0.77395381   0.089465498   0.93393927   0   0
#>   1.01360655   0.717402455   2.11914590   0   0
#>   0.39842033   0.255535362   0.47209263   0   0
#>  21.76314446   1.662309762  23.16159471   0   1
#>   0.32178397   0.165152729   0.31226840   0   0
#>   0.44182769   0.242488562   0.46547284   0   0
#>   0.40807389   0.739419599   0.40994367   0   0
#>   0.57247899   0.711170649   0.58271230   0   0
#>   0.54724057   0.787394209   0.53714317   0   0
#>   0.57983231   0.711170649   0.58271230   0   0
#>   0.02719314   0.194299848   0.11988073   0   0
#>  -0.39141835   0.129072863   0.05357031   0   0
#>   0.46877814   0.787394209   0.53714317   0   0
#>   0.36018100   0.739419599   0.40994367   0   0
#>   0.11988073   0.220992603   0.17961004   0   0
#>   0.17961004   0.191106953   0.20181627   0   0
#>   0.46547284   0.114285745   0.54277163   0   0
#>   0.05086777   0.549929505   0.10333877   0   0
#>   0.10333877   0.561273185   0.12694579   0   0
#>   0.12694579   0.549545682   0.15708810   0   0
#>   0.15708810   0.529180107   0.19029221   0   0
#>   0.19029221   0.522610528   0.19728971   0   0
#>   0.19728971   0.456702715   0.16692419   0   0
#>   0.16692419   0.417561601   0.12317554   0   0
#>   0.12317554   0.411365707   0.11005809   0   0
#>   0.11005809   0.509446587   0.05086777   0   0
#>   0.85825200   0.724357169   0.97038077   0   0
#>   0.97038077   0.676787720   1.05431542   0   0
#>   1.05431542   0.602255013   0.83142671   0   0
#>   0.83142671   0.724671783   0.76950468   0   0
#>   0.76950468   0.746055472   0.85825200   0   0
#>   0.63359360   0.943350662   0.31571543   0   0
#>   0.31571543   0.892737920   0.27834075   0   0
#>   0.27834075   0.897449731   0.25317619   0   0
#>   0.25317619   0.973617637   0.12661461   0   0
#>   0.12661461   6.530155729  -0.56833208   0   1
#>   0.74171561   0.936906076   0.84646453   0   0
#>   0.84646453   0.931076612   0.84947219   0   0
#>   0.84947219   0.832431816   0.80827545   0   0
#>   0.80827545   0.764713986   0.75292535   0   0
#>   0.75292535   0.805446004   0.72440156   0   0
#>   0.72440156   0.931937546   0.68752389   0   0
#>   0.68752389   1.053059544   0.74171561   0   0
#>   0.64181742   0.346665629   0.65419526   0   0
#>   0.65419526   0.305631018   0.62845540   0   0
#>   0.62845540   0.304496205   0.62505668   0   0
#>   0.62505668   0.296297226   0.56188120   0   0
#>   0.56188120   0.297178336   0.55860335   0   0
#>   0.55860335   0.342406279   0.54805334   0   0
#>   0.54805334   0.399294908   0.57108336   0   0
#>   0.57108336   0.377269237   0.64181742   0   0
#>   0.79406324   0.544474639   0.78609274   0   0
#>   0.78609274   0.656918651   0.72538585   0   0
#>   0.72538585   0.720420485   0.75743708   0   0
#>   0.75743708   0.725851755   0.76558873   0   0
#>   0.17876064   0.340854069   0.08927012   0   0
#>   0.08927012   0.352463779   0.08447135   0   0
#>   0.17274282   0.327024262   0.18932175   0   0
#>   0.18932175   0.302299835   0.17876064   0   0
#>   0.47044620   0.020599182   0.40794098   0   0
#>   0.40794098   0.059610391   0.37789508   0   0
#>   0.37789508   0.136128885   0.42925470   0   0
#>   0.42925470   0.099578057   0.47044620   0   0
#>   0.70057093   0.574751586   0.61992925   0   0
#>   0.61992925   0.637155055   0.60705204   0   0
#>   0.60705204   0.653363004   0.72087136   0   0
#>   0.72087136   0.599603988   0.70057093   0   0
#>   0.12835049   0.026112288   0.15647325   0   0
#>   0.15647325   0.073179945   0.21943049   0   0
#>   0.21943049   0.048617074   0.26068889   0   0
#>   0.26068889  -1.821157119   0.56968718   0   0
#>   0.56968718  -2.640255351   0.60467423   0   0
#>   0.60467423 -45.330654154   0.98417909   0   0
#>   0.98417909  -0.320423293   0.12835049   0   0
#>   0.46863451   0.909338157   0.44839485   0   0
#>   0.44839485   1.001063084   0.49071321   0   0
#>   0.49071321   1.061716220   0.52375291   0   0
#>   0.52375291   0.922606247   0.53279757   0   0
#>   0.53279757   0.879389740   0.46863451   0   0
#>   0.29724128   0.723619175   0.24480973   0   0
#>   0.24480973   0.723818653   0.19906141   0   0
#>   0.19906141   0.774718286   0.15249117   0   0
#>   0.15249117   0.882135455   0.28054011   0   0
#>   0.28054011   0.832989805   0.29724128   0   0
#>   0.62888767   0.506858300   0.46230122   0   0
#>   0.46230122   0.543071484   0.44332990   0   0
#>   0.44332990   0.567030190   0.48915890   0   0
#>   0.48915890   0.573530613   0.61974818   0   0
#>   0.61974818   0.516865704   0.62888767   0   0
#>   0.08084301   0.104738780   0.07015556   0   0
#>   0.07015556   0.097624938   0.20937936   0   0
#>   0.20937936   0.073800004   0.21925575   0   0
#>   0.74929521   0.142562212   0.76317208   0   0
#>   0.76317208   0.134911286   0.76310208   0   0
#>   0.76310208   0.005241968   0.69602512   0   0
#>   0.69602512   0.123194672   0.59472698   0   0
#>   0.59472698   0.140962009   0.61253496   0   0
#>   0.61253496   0.154524844   0.65709192   0   0
#>   0.65709192   0.161229420   0.74929521   0   0
#>  -0.02414489   0.476729401  -0.04414038   0   0
#>  -0.04414038   0.611098047   0.01138247   0   0
#>   0.25376171   0.675784605   0.28643621   0   0
#>   0.08708354   0.815847578   0.05001599   0   0
#>   0.05001599   0.843656462   0.01427949   0   0
#>   0.50782212  -1.203617743   0.56322540   0   0
#>   0.46749576   0.434000050   0.46048041   0   0
#>   0.46048041   0.389265678   0.40817177   0   0
#>   0.40817177   0.451478356   0.34605133   0   0
#>   0.34605133   0.561224714   0.36171875   0   0
#>   0.36171875   0.554246272   0.42767230   0   0
#>   0.06461354   0.161231566  -0.74811734   0   0
#>  -0.74811734   0.005757352  -3.75714489   0   1
#>   0.85448182   0.088973970   0.93782210   0   0
#>   0.93782210   0.113285435   0.97971501   0   0
#>   0.97971501  -0.885344483   5.57506594   0   1
#>   0.61439443   0.501068007   0.63804202   0   0
#>   0.63804202   0.486642772   0.69494064   0   0
#>   0.69494064   0.365231329   0.68830894   0   0
#>   0.33012331   0.562432940   0.35946873   0   0
#>   0.35946873   0.517089164   0.21125994   0   0
#>   0.85998738   0.843607512   0.91203702   0   0
#>   0.85406712   0.925741814   0.85395272   0   0
#>   0.85395272   0.921153038   0.85998738   0   0
#>   0.77293183   0.402856907   0.83214982   0   0
#>   0.83214982   0.375486676   0.87488329   0   0
#>   0.87488329   0.232652964   0.78716118   0   0
#>   0.78716118   0.213771531   0.75246498   0   0
#>   0.75246498   0.241374378   0.74028009   0   0
#>   0.74028009   0.357330470   0.77293183   0   0
#>   0.45705483   0.684815260   0.57530052   0   0
#>   1.37853497   0.428691856   0.94598435   0   0
#>   0.94598435   0.481382509   0.93295860   0   0
#>   0.93295860   0.534363308   0.93958288   0   0
#>   0.93958288   0.663887451   1.11352889   0   0
#>   1.11352889   0.659659858   1.37853497   0   0
#>   0.35860620   0.217928962   0.35350879   0   0
#>   0.35350879   0.276550409   0.25824892   0   0
#>   0.25824892   0.343664646   0.33935300   0   0
#>   0.33935300   0.220745622   0.35860620   0   0
#>   0.91181574   0.247956296   0.92357704   0   0
#>   0.92357704   0.204451319   0.92408403   0   0
#>   0.92408403   0.143266659   0.77395381   0   0
#>   0.81719814   0.089465498   0.93393927   0   0
#>   0.39842033   0.343883852   0.33945974   0   0
#>   0.33945974   0.409106925   0.26281988   0   0
#>   0.26281988   0.419846501   0.26277308   0   0
#>   1.27866512   0.363587718   0.89094505   0   0
#>   2.11914590   1.593711862  21.76314446   0   0
#>  21.76314446   0.327342303   1.27866512   0   0
#>   0.79130357   0.481134024   0.80454686   0   0
#>   0.33610171   0.164265944   0.32178397   0   0
#>   0.44182769   0.147632026   0.42541491   0   0
#>   0.42541491   0.125184451   0.33610171   0   0
#>   0.57247899   0.784456200   0.54724057   0   0
#>   0.54724057   0.887054420   0.62829467   0   0
#>   0.02719314   0.174139299  -0.39141835   0   0
#>   0.46877814   0.798752312   0.36018100   0   0
#>   0.11988073   0.129072863   0.05357031   0   0
#>   0.19122176   0.220992603   0.17961004   0   0
#>   0.47209263   0.242488562   0.46547284   0   0
#>   0.20181627   0.165152729   0.31226840   0   0
#>   0.53714317   0.739419599   0.40994367   0   0
#>   0.63359360   6.167001719   0.65154226   0   1
#>  -0.03564234   1.116019986  -8.56347315   0   1
#>   0.06582752  -7.764649709 -26.11108706   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.445201599 0.116101332
#>   [2,] 0.672382896 0.850478630
#>   [3,] 1.000000000 0.213299094
#>   [4,] 0.836902255 0.782542103
#>   [5,] 0.385945992 0.605049540
#>   [6,] 0.628460612 0.763646458
#>   [7,] 0.394541306 0.140030206
#>   [8,] 0.407467630 0.149861219
#>   [9,] 0.100672358 0.426829551
#>  [10,] 0.625147749 0.686267091
#>  [11,] 0.003240751 0.207446833
#>  [12,] 0.936645428 0.510241520
#>  [13,] 0.793496157 0.198005879
#>  [14,] 0.544382087 0.541569086
#>  [15,] 0.081474201 0.148958476
#>  [16,] 0.082581794 0.668526689
#>  [17,] 0.588845603 0.005275755
#>  [18,] 0.591025612 0.000000000
#>  [19,] 0.633377521 0.231518111
#>  [20,] 0.654153003 0.197398296
#>  [21,] 0.220637632 0.626503431
#>  [22,] 0.884669450 0.143890305
#>  [23,] 0.022783088 0.325699131
#>  [24,] 0.463052611 0.386324318
#>  [25,] 0.366540400 0.009435556
#>  [26,] 0.009884877 0.954842778
#>  [27,] 0.409067002 0.662215082
#>  [28,] 0.608515825 0.269428294
#>  [29,] 0.817309464 0.878523713
#>  [30,] 0.292731381 0.806074716
#>  [31,] 0.622992310 0.500472855
#>  [32,] 0.479053095 0.947387478
#>  [33,] 0.278721766 0.321311078
#>  [34,] 0.044107130 0.623726352
#>  [35,] 0.472598314 0.921277142
#>  [36,] 0.215910105 0.847880325
#>  [37,] 0.864661331 0.949070318
#>  [38,] 0.051667859 0.866037408
#>  [39,] 0.411222588 0.334417334
#>  [40,] 0.974112745 0.429031214
#>  [41,] 0.401843244 0.988615034
#>  [42,] 0.227365315 0.657998734
#>  [43,] 0.609317403 0.728188553
#>  [44,] 0.405964106 0.719022074
#>  [45,] 0.158316665 0.387842903
#>  [46,] 0.410598509 0.191206712
#>  [47,] 0.652073691 0.406324342
#>  [48,] 0.413333737 0.537396483
#>  [49,] 0.800728639 0.658465494
#>  [50,] 0.421788754 0.077305250
#>  [51,] 0.931043665 0.965007655
#>  [52,] 0.142991089 0.877598027
#>  [53,] 0.000000000 0.037007413
#>  [54,] 0.594771147 0.539060839
#>  [55,] 0.343033909 0.015417869
#>  [56,] 0.862137303 0.399992607
#>  [57,] 0.307892188 0.979832164
#>  [58,] 0.473859348 0.310625804
#>  [59,] 0.226560606 0.045952323
#>  [60,] 0.939447983 0.553345795
#>  [61,] 0.267086300 0.085120146
#>  [62,] 0.982533878 0.426001094
#>  [63,] 0.479093896 0.545491166
#>  [64,] 0.287348654 0.376388089
#>  [65,] 0.217682858 1.000000000
#>  [66,] 0.853927250 0.953367776
#>  [67,] 0.128958828 0.263506224
#>  [68,] 0.391000472 0.606623440
#>  [69,] 0.142754792 0.464170945
#>  [70,] 0.557693365 0.888830151
#>  [71,] 0.117050147 0.275202380
#>  [72,] 0.219276071 0.284729119
#>  [73,] 0.818037812 0.432504516
#>  [74,] 0.731047676 0.129751888
#>  [75,] 0.808102576 0.817776937
#>  [76,] 0.625330802 0.029545397
#>  [77,] 0.990190287 0.952519390
#>  [78,] 0.011759334 0.102478752
#>  [79,] 0.035091991 0.026561249
#>  [80,] 0.276691518 0.832191656
#>  [81,] 0.670797237 0.679766573
#>  [82,] 0.057379641 0.481532350
#>  [83,] 0.896427428 0.537330028
#>  [84,] 0.338092592 0.272181274
#>  [85,] 0.816273847 0.178898223
#>  [86,] 0.695281570 0.478803304
#>  [87,] 0.119850067 0.150919346
#>  [88,] 0.120452868 0.397359418
#>  [89,] 0.977838344 0.036102643
#>  [90,] 0.761565677 0.674961341
#>  [91,] 0.197086987 0.554282839
#>  [92,] 0.319517027 0.710950898
#>  [93,] 0.570439810 0.703127153
#>  [94,] 0.356234303 0.477675374
#>  [95,] 0.080860026 0.856704108
#>  [96,] 0.027484852 0.775036926
#>  [97,] 0.707941971 0.006367211
#>  [98,] 0.730891059 0.328598232
#>  [99,] 0.482677402 0.918400379
#> [100,] 0.817272621 0.829544889
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.4452016,0.1161013) [9]: 7 8 17 19 20 28 46 50 76 
#> 2: (0.6723829,0.8504786) [5]: 6 29 66 70 75 
#> 3: (1,0.2132991) [6]: 22 56 62 77 85 89 
#> 4: (0.8369023,0.7825421) [7]: 49 51 60 75 77 90 100 
#> 5: (0.385946,0.6050495) [8]: 21 27 42 48 68 91 92 94 
#> 6: (0.6284606,0.7636465) [7]: 2 43 70 75 81 90 93 
#> 7: (0.3945413,0.1400302) [7]: 1 8 46 50 55 61 84 
#> 8: (0.4074676,0.1498612) [3]: 1 7 46 
#> 9: (0.1006724,0.4268296) [4]: 23 69 82 88 
#> 10: (0.6251477,0.6862671) [4]: 43 54 81 93 
#> 11: (0.003240751,0.2074468) [7]: 15 23 26 53 71 78 96 
#> 12: (0.9366454,0.5102415) [5]: 40 56 60 62 83 
#> 13: (0.7934962,0.1980059) [5]: 20 56 74 85 98 
#> 14: (0.5443821,0.5415691) [5]: 24 31 54 63 93 
#> 15: (0.0814742,0.1489585) [6]: 11 67 71 78 79 87 
#> 16: (0.08258179,0.6685267) [7]: 21 34 36 42 91 95 96 
#> 17: (0.5888456,0.005275755) [5]: 1 18 25 50 76 
#> 18: (0.5910256,0) [4]: 17 25 76 97 
#> 19: (0.6333775,0.2315181) [4]: 1 20 28 98 
#> 20: (0.654153,0.1973983) [6]: 1 13 19 74 76 98 
#> 21: (0.2206376,0.6265034) [4]: 5 16 42 91 
#> 22: (0.8846694,0.1438903) [5]: 3 74 85 89 97 
#> 23: (0.02278309,0.3256991) [7]: 9 11 34 71 82 88 96 
#> 24: (0.4630526,0.3863243) [8]: 14 31 39 47 48 58 63 94 
#> 25: (0.3665404,0.009435556) [5]: 17 18 50 55 79 
#> 26: (0.009884877,0.9548428) [6]: 11 38 52 53 65 96 
#> 27: (0.409067,0.6622151) [6]: 5 44 63 68 92 93 
#> 28: (0.6085158,0.2694283) [6]: 1 19 46 47 58 98 
#> 29: (0.8173095,0.8785237) [6]: 2 37 51 66 75 100 
#> 30: (0.2927314,0.8060747) [6]: 35 36 42 44 80 92 
#> 31: (0.6229923,0.5004729) [5]: 14 24 47 54 86 
#> 32: (0.4790531,0.9473875) [5]: 35 41 66 70 99 
#> 33: (0.2787218,0.3213111) [4]: 45 64 72 84 
#> 34: (0.04410713,0.6237264) [5]: 16 23 82 91 96 
#> 35: (0.4725983,0.9212771) [6]: 30 32 41 44 80 99 
#> 36: (0.2159101,0.8478803) [8]: 16 30 42 52 57 65 80 95 
#> 37: (0.8646613,0.9490703) [3]: 29 51 66 
#> 38: (0.05166786,0.8660374) [4]: 26 52 95 96 
#> 39: (0.4112226,0.3344173) [6]: 24 46 58 64 84 94 
#> 40: (0.9741127,0.4290312) [3]: 12 56 62 
#> 41: (0.4018432,0.988615) [7]: 32 35 51 57 65 66 80 
#> 42: (0.2273653,0.6579987) [6]: 5 16 21 30 36 92 
#> 43: (0.6093174,0.7281886) [4]: 6 10 81 93 
#> 44: (0.4059641,0.7190221) [7]: 27 30 35 70 92 93 99 
#> 45: (0.1583167,0.3878429) [6]: 33 64 69 71 72 88 
#> 46: (0.4105985,0.1912067) [7]: 1 7 8 28 39 58 84 
#> 47: (0.6520737,0.4063243) [6]: 24 28 31 58 86 98 
#> 48: (0.4133337,0.5373965) [5]: 5 24 63 68 94 
#> 49: (0.8007286,0.6584655) [5]: 4 60 83 86 90 
#> 50: (0.4217888,0.07730525) [5]: 1 7 17 25 55 
#> 51: (0.9310437,0.9650077) [8]: 4 29 37 41 65 66 77 100 
#> 52: (0.1429911,0.877598) [5]: 26 36 38 65 95 
#> 53: (0,0.03700741) [4]: 11 26 78 79 
#> 54: (0.5947711,0.5390608) [6]: 10 14 31 81 86 93 
#> 55: (0.3430339,0.01541787) [6]: 7 25 50 59 61 79 
#> 56: (0.8621373,0.3999926) [9]: 3 12 13 40 62 73 83 85 98 
#> 57: (0.3078922,0.9798322) [4]: 36 41 65 80 
#> 58: (0.4738593,0.3106258) [5]: 24 28 39 46 47 
#> 59: (0.2265606,0.04595232) [4]: 55 61 79 87 
#> 60: (0.939448,0.5533458) [6]: 4 12 49 62 77 83 
#> 61: (0.2670863,0.08512015) [6]: 7 55 59 72 84 87 
#> 62: (0.9825339,0.4260011) [6]: 3 12 40 56 60 77 
#> 63: (0.4790939,0.5454912) [6]: 14 24 27 48 68 93 
#> 64: (0.2873487,0.3763881) [7]: 33 39 45 69 84 91 94 
#> 65: (0.2176829,1) [6]: 26 36 41 51 52 57 
#> 66: (0.8539273,0.9533678) [7]: 2 29 32 37 41 51 70 
#> 67: (0.1289588,0.2635062) [4]: 15 71 72 87 
#> 68: (0.3910005,0.6066234) [4]: 5 27 48 63 
#> 69: (0.1427548,0.4641709) [6]: 9 45 64 82 88 91 
#> 70: (0.5576934,0.8888302) [7]: 2 6 32 44 66 93 99 
#> 71: (0.1170501,0.2752024) [7]: 11 15 23 45 67 72 88 
#> 72: (0.2192761,0.2847291) [7]: 33 45 61 67 71 84 87 
#> 73: (0.8180378,0.4325045) [4]: 56 83 86 98 
#> 74: (0.7310477,0.1297519) [6]: 13 20 22 76 85 97 
#> 75: (0.8081026,0.8177769) [6]: 2 4 6 29 90 100 
#> 76: (0.6253308,0.0295454) [6]: 1 17 18 20 74 97 
#> 77: (0.9901903,0.9525194) [5]: 3 4 51 60 62 
#> 78: (0.01175933,0.1024788) [4]: 11 15 53 79 
#> 79: (0.03509199,0.02656125) [7]: 15 25 53 55 59 78 87 
#> 80: (0.2766915,0.8321917) [5]: 30 35 36 41 57 
#> 81: (0.6707972,0.6797666) [6]: 6 10 43 54 86 90 
#> 82: (0.05737964,0.4815323) [5]: 9 23 34 69 91 
#> 83: (0.8964274,0.53733) [6]: 12 49 56 60 73 86 
#> 84: (0.3380926,0.2721813) [7]: 7 33 39 46 61 64 72 
#> 85: (0.8162738,0.1788982) [5]: 3 13 22 56 74 
#> 86: (0.6952816,0.4788033) [9]: 31 47 49 54 73 81 83 90 98 
#> 87: (0.1198501,0.1509193) [6]: 15 59 61 67 72 79 
#> 88: (0.1204529,0.3973594) [5]: 9 23 45 69 71 
#> 89: (0.9778383,0.03610264) [3]: 3 22 97 
#> 90: (0.7615657,0.6749613) [6]: 4 6 49 75 81 86 
#> 91: (0.197087,0.5542828) [8]: 5 16 21 34 64 69 82 94 
#> 92: (0.319517,0.7109509) [5]: 5 27 30 42 44 
#> 93: (0.5704398,0.7031272) [9]: 6 10 14 27 43 44 54 63 70 
#> 94: (0.3562343,0.4776754) [6]: 5 24 39 48 64 91 
#> 95: (0.08086003,0.8567041) [5]: 16 36 38 52 96 
#> 96: (0.02748485,0.7750369) [7]: 11 16 23 26 34 38 95 
#> 97: (0.707942,0.006367211) [5]: 18 22 74 76 89 
#> 98: (0.7308911,0.3285982) [8]: 13 19 20 28 47 56 73 86 
#> 99: (0.4826774,0.9184004) [4]: 32 35 44 70 
#> 100: (0.8172726,0.8295449) [4]: 4 29 51 75 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  3 18 25 26 51 53 65 77 79 89 97 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.181923717 0.043819919 0.170309469 0.039947246 0.045335260 0.212814191 
#>                                                                         
#> 0.041463865 0.016240033 0.068387452 0.061440520 0.083321642 0.303182486 
#>                                                                         
#> 0.143801808 0.187620213 0.526573961 0.123173766 0.225801203 0.076834321 
#>                                                                         
#> 0.142473235 0.402385810 0.060450638 0.176901135 0.014918927 0.150206450 
#>                                                                         
#> 0.042495301 0.174049439 0.058453658 0.101952821 0.106282030 0.032205835 
#>                                                                         
#> 0.075963456 0.176625514 0.082625396 0.072740266 0.186139935 0.046255174 
#>                                                                         
#> 0.078258917 0.090895544 0.200084380 0.103084657 0.100161737 0.108692170 
#>                                                                         
#> 0.087122527 0.159627421 0.121138692 0.070435473 0.165860692 0.159931454 
#>                                                                         
#> 0.066519005 0.083788588 0.131160861 0.106940171 0.449362394 0.180665170 
#>                                                                         
#> 0.917888592 0.141553452 0.115679472 0.008949698 0.134436108 0.045905042 
#>                                                                         
#> 0.149407943 0.151993923 0.102414976 0.098381210 0.182189956 0.159973149 
#>                                                                         
#> 0.196495791 0.047806553 0.079422164 0.150495408 0.112954746 0.016691817 
#>                                                                         
#> 0.190226813 0.135338032 0.097562481 0.152220860 0.168003684 0.087513862 
#>                                                                         
#> 0.224683422 0.045274457 0.085801128 0.135929318 0.145661932 0.271529460 
#>                                                                         
#> 0.125529487 0.142812093 0.122204420 0.066256490 0.153448941 0.067002990 
#>                                                                         
#> 0.202283371 0.098537654 0.308143473 0.094158937 0.092052052 0.143391434 
#>                                                                         
#> 0.107226892 0.165241961 0.086823045 0.110695645 0.135140125 0.068268683 
#>                                                                         
#> 0.011562380 0.212948892 0.215210819 0.062773518 0.084380807 0.117189231 
#>                                                                         
#> 0.453455971 0.097676361 0.010481592 0.080633698 0.129539442 0.119785231 
#>                                                                         
#> 0.119477254 0.115905166 0.157514368 0.200327801 0.213627336 0.150901297 
#>                                                                         
#> 0.092436274 0.065550171 0.077989924 0.122492232 0.714218528 0.102668877 
#>                                                                         
#> 0.169154411 0.154368388 0.207054482 0.209487550 0.202449298 0.056360005 
#>                                                                         
#> 0.192448031 0.130894308 0.135613688 0.161270094 0.166705085 0.105224242 
#>                                                                         
#> 0.181427419 0.199846090 0.224209468 0.220751468 0.224010477 0.082693360 
#>                                                                         
#> 0.050631832 0.056027282 0.045313323 0.147615165 0.208673258 0.120931914 
#>                                                                         
#> 0.097308752 0.139603840 0.213417915 0.232078846 0.186919018 0.134605780 
#>                                                                         
#> 0.178576934 0.228887524 0.205319953 0.050937063 0.045507308 0.131336287 
#>                                                                         
#> 0.129242158 0.251090768 0.061664254 0.125011619 0.167186867 0.166694701 
#>                                                                         
#> 0.195563243 0.130793588 0.072986470 0.005293857 0.083839130 0.040295476 
#>                                                                         
#> 0.093958634 0.159943766 0.138780018 0.134837853 0.053636447 0.143702338 
#>                                                                         
#> 0.069761419 0.127648346 0.035493060 0.056261096 0.057247012 0.150307746 
#>                                                                         
#> 0.046110004 0.044810812 0.105313173 0.097679889 0.132451474 0.119856202 
#>                                                                         
#> 0.568107640 0.747425477 0.170470227 0.133064959 0.089436646 0.095928094 
#>                                                                         
#> 0.043195287 0.048489945 0.144823206 0.139344478 0.092511728 0.029730887 
#>                                                                         
#> 0.213331305 0.065405891 0.175258158 0.088704382 0.050451449 0.163646001 
#>                                                                         
#> 0.130890759 0.038425929 0.123999905 0.223481912 0.188185296 0.119917069 
#>                                                                         
#> 0.059053958 0.161750105 0.144309975 0.145165788 0.222344118 0.005708418 
#>                                                                         
#> 0.137598837 0.154271009 0.223931246 0.298789121 0.159043282 0.140553074 
#>                                                                         
#> 0.073353161 0.076466009 0.190076221 0.024255787 0.331890550 0.098143830 
#>                                                                         
#> 0.153896118 0.212648009 0.136118472 0.166478168 0.056891672 0.143658678 
#>                                                                         
#> 0.140817618 0.084964832 0.048978838 0.142880862 0.142827912 0.213597080 
#>                                                                         
#> 0.030649173 0.087459805 0.161861728 0.098823132 0.075467255 0.087527552 
#>                                                                         
#> 0.026896353 0.029212797 0.098047196 0.374921854 0.137564065 0.069799931 
#>                                                                         
#> 0.077062525 0.055748545 0.160847727 0.152130004 0.078742141 0.030647889 
#>                                                                         
#> 0.130791039 0.096027723 0.143211982 0.094360690 0.529726716 0.184511965 
#>                                                                         
#> 0.227720336 0.119961714 0.077898278 0.039041403 0.208320875 0.154376355 
#>                                                                         
#> 0.120409231 0.054788587 0.149683718 0.205254866 0.199483610 0.092777237 
#>                                                 
#> 0.131197076 0.739285382 0.117089608 0.036613797 
#> 
#> $alpha
#> [1] 0.3007843
#> 
#> attr(,"class")
#> [1] "scree"