Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1         x2          y2         mx1
#>    73   35 0.982097281 0.378060629 0.92268857 0.421914805  0.90357064
#>     1   35 0.992776850 0.321067002 0.92268857 0.421914805  0.87555563
#>     5   35 0.872899924 0.196996766 0.92268857 0.421914805  0.86552631
#>     5   92 0.872899924 0.196996766 0.74571322 0.314643524  0.81825473
#>    92   35 0.745713221 0.314643524 0.92268857 0.421914805  0.83379306
#>    35   78 0.922688570 0.421914805 0.95926174 0.505332550  0.88665631
#>    38   78 0.959240659 0.443250718 0.95926174 0.505332550  0.91661017
#>    38   73 0.959240659 0.443250718 0.98209728 0.378060629  1.15198046
#>    83   78 0.974832334 0.769953479 0.95926174 0.505332550  1.79217313
#>     1    5 0.992776850 0.321067002 0.87289992 0.196996766  0.96445800
#>    34   71 0.885444693 0.112809722 0.93577739 0.136006320  0.89543970
#>    59   35 0.796816961 0.465596910 0.92268857 0.421914805  0.87071701
#>    20   34 0.846059733 0.182398554 0.88544469 0.112809722  0.87846187
#>    13   34 0.826970240 0.125159586 0.88544469 0.112809722  0.86177673
#>    92   51 0.745713221 0.314643524 0.66342646 0.179189620  0.75411009
#>    92   59 0.745713221 0.314643524 0.79681696 0.465596910  0.74139872
#>    59    7 0.796816961 0.465596910 0.83327503 0.539804640  0.77688565
#>    51   13 0.663426465 0.179189620 0.82697024 0.125159586  0.75473857
#>    70   51 0.548777559 0.295675201 0.66342646 0.179189620  0.64964121
#>    67   12 0.879380200 0.622385568 0.75731361 0.668558547  0.80486394
#>     7   78 0.833275033 0.539804640 0.95926174 0.505332550  0.90486726
#>    67   78 0.879380200 0.622385568 0.95926174 0.505332550  0.92175549
#>    13  100 0.826970240 0.125159586 0.65452139 0.126461400  0.74059799
#>     6   97 0.916822256 0.009644253 1.00000000 0.023726228  0.94949571
#>     6   34 0.916822256 0.009644253 0.88544469 0.112809722  0.84008763
#>    97    1 1.000000000 0.023726228 0.99277685 0.321067002  1.13557017
#>   100   51 0.654521390 0.126461400 0.66342646 0.179189620  0.60855126
#>    70   92 0.548777559 0.295675201 0.74571322 0.314643524  0.64214580
#>     3   29 0.634657395 0.570469451 0.53370977 0.530415197  0.62054867
#>     7   12 0.833275033 0.539804640 0.75731361 0.668558547  0.74790945
#>    12    3 0.757313613 0.668558547 0.63465740 0.570469451  0.72826168
#>     6   13 0.916822256 0.009644253 0.82697024 0.125159586  0.82160171
#>    57   58 0.910492190 0.635537914 0.86639938 0.706390509  0.87952648
#>    92   98 0.745713221 0.314643524 0.72998078 0.485163452  0.64678827
#>    90   12 0.809734986 0.707398676 0.75731361 0.668558547  0.81722191
#>    70   77 0.548777559 0.295675201 0.52445198 0.310656582  0.52486613
#>    70   55 0.548777559 0.295675201 0.53065105 0.344538924  0.54922003
#>    58   90 0.866399379 0.706390509 0.80973499 0.707398676  0.83721282
#>    56   62 0.625335964 0.044569267 0.72500109 0.019192461  0.68475707
#>    70   80 0.548777559 0.295675201 0.51160663 0.307114877  0.52067165
#>    77   55 0.524451981 0.310656582 0.53065105 0.344538924  0.51209010
#>    21   80 0.492097910 0.295143780 0.51160663 0.307114877  0.47106357
#>     8   56 0.550912586 0.164104608 0.62533596 0.044569267  0.58775747
#>     8   70 0.550912586 0.164104608 0.54877756 0.295675201  0.52105642
#>    29   55 0.533709767 0.530415197 0.53065105 0.344538924  0.61531883
#>    12   61 0.757313613 0.668558547 0.74676444 0.717975179  0.66716234
#>    39   45 0.852868618 0.723629371 0.87396174 0.849163641  0.80473892
#>     4   21 0.395451126 0.316010645 0.49209791 0.295143780  0.45615904
#>    31   29 0.580050078 0.699275190 0.53370977 0.530415197  0.55884346
#>     8   21 0.550912586 0.164104608 0.49209791 0.295143780  0.47284597
#>    39   83 0.852868618 0.723629371 0.97483233 0.769953479  0.90330185
#>    39   58 0.852868618 0.723629371 0.86639938 0.706390509  0.91069840
#>    57   78 0.910492190 0.635537914 0.95926174 0.505332550  1.09431568
#>     1   83 0.992776850 0.321067002 0.97483233 0.769953479  2.31771796
#>    24   69 0.549833803 0.022247725 0.49350707 0.017438755  0.61776956
#>    56   30 0.625335964 0.044569267 0.51743547 0.130233751  0.57155813
#>    39   88 0.852868618 0.723629371 0.92665936 0.842330696  0.89199234
#>    53   30 0.437481850 0.119640392 0.51743547 0.130233751  0.46660888
#>     4   63 0.395451126 0.316010645 0.37245752 0.229915644  0.43385540
#>    29   43 0.533709767 0.530415197 0.37455158 0.398063891  0.47561148
#>    43    4 0.374551580 0.398063891 0.39545113 0.316010645  0.45371116
#>    90   39 0.809734986 0.707398676 0.85286862 0.723629371  0.80221899
#>    11   61 0.720893137 0.819939784 0.74676444 0.717975179  0.79009890
#>    62   68 0.725001094 0.019192461 0.45177023 0.013583659  0.62996957
#>    69   30 0.493507066 0.017438755 0.51743547 0.130233751  0.49212577
#>    31   61 0.580050078 0.699275190 0.74676444 0.717975179  0.65877593
#>    43   72 0.374551580 0.398063891 0.36815604 0.358331031  0.43062266
#>    31   93 0.580050078 0.699275190 0.50377114 0.701475422  0.53958047
#>    72   28 0.368156039 0.358331031 0.36753678 0.321418848  0.38553609
#>    53   63 0.437481850 0.119640392 0.37245752 0.229915644  0.39154149
#>    96   43 0.417308439 0.644903039 0.37455158 0.398063891  0.40831180
#>    32   93 0.542744739 0.784879948 0.50377114 0.701475422  0.54288074
#>    72   40 0.368156039 0.358331031 0.35155022 0.375936763  0.34701086
#>    43   66 0.374551580 0.398063891 0.32092273 0.426792087  0.34446531
#>    69   68 0.493507066 0.017438755 0.45177023 0.013583659  0.62148302
#>    69   87 0.493507066 0.017438755 0.43532119 0.059839520  0.46981750
#>    91   17 0.294803409 0.183402073 0.37982092 0.102030715  0.35860526
#>    87   53 0.435321185 0.059839520 0.43748185 0.119640392  0.41485625
#>     4   28 0.395451126 0.316010645 0.36753678 0.321418848  0.37644672
#>    88   94 0.926659364 0.842330696 0.97765905 0.895067332  0.91915300
#>    45   88 0.873961741 0.849163641 0.92665936 0.842330696  0.90699425
#>    88   83 0.926659364 0.842330696 0.97483233 0.769953479  0.98989931
#>    87   17 0.435321185 0.059839520 0.37982092 0.102030715  0.39781495
#>    82   89 0.330717817 0.368899455 0.31329101 0.262481584  0.30642541
#>    79   28 0.348628312 0.281157443 0.36753678 0.321418848  0.31727010
#>    93   96 0.503771139 0.701475422 0.41730844 0.644903039  0.50012600
#>    66   82 0.320922725 0.426792087 0.33071782 0.368899455  0.33218086
#>    11   32 0.720893137 0.819939784 0.54274474 0.784879948  0.63719202
#>    11   50 0.720893137 0.819939784 0.85832283 0.934771631  0.71080962
#>    96   66 0.417308439 0.644903039 0.32092273 0.426792087  0.40528823
#>    91   63 0.294803409 0.183402073 0.37245752 0.229915644  0.32710626
#>    27   89 0.266352224 0.245091357 0.31329101 0.262481584  0.26512935
#>    11   45 0.720893137 0.819939784 0.87396174 0.849163641  0.78915085
#>    68   87 0.451770231 0.013583659 0.43532119 0.059839520  0.39603680
#>    17   49 0.379820924 0.102030715 0.30970296 0.073119839  0.34086367
#>    76   36 0.602977744 0.864454291 0.45319144 0.847828033  0.52779143
#>    42   86 0.895959051 0.954616364 0.92228447 1.000000000  0.83827558
#>    50   45 0.858322825 0.934771631 0.87396174 0.849163641  0.90153337
#>    63   79 0.372457518 0.229915644 0.34862831 0.281157443  0.34370383
#>    27   91 0.266352224 0.245091357 0.29480341 0.183402073  0.30098383
#>    81   84 0.319447507 0.059687468 0.33581249 0.000000000  0.34162197
#>    49   81 0.309702959 0.073119839 0.31944751 0.059687468  0.33800794
#>    91   49 0.294803409 0.183402073 0.30970296 0.073119839  0.32662866
#>    27   14 0.266352224 0.245091357 0.22671998 0.213306947  0.26448139
#>     6   62 0.916822256 0.009644253 0.72500109 0.019192461  0.67486171
#>    84   68 0.335812490 0.000000000 0.45177023 0.013583659  0.39294485
#>    86   76 0.922284468 1.000000000 0.60297774 0.864454291  0.64602283
#>    42   94 0.895959051 0.954616364 0.97765905 0.895067332  0.95545922
#>    93   36 0.503771139 0.701475422 0.45319144 0.847828033  0.45605586
#>    14   19 0.226719984 0.213306947 0.21818892 0.220518257  0.20450538
#>    19   27 0.218188919 0.220518257 0.26635222 0.245091357  0.22357799
#>    91   16 0.294803409 0.183402073 0.22348361 0.174605602  0.26145993
#>    37   48 0.387718645 0.810344984 0.25037057 0.769201342  0.34095970
#>    14   16 0.226719984 0.213306947 0.22348361 0.174605602  0.25763485
#>    49   95 0.309702959 0.073119839 0.22777890 0.137536897  0.23760396
#>    18   16 0.161474274 0.211363552 0.22348361 0.174605602  0.19331066
#>    52   18 0.191091307 0.306418229 0.16147427 0.211363552  0.18324911
#>    15   66 0.247902441 0.663838205 0.32092273 0.426792087  0.32169972
#>    66   26 0.320922725 0.426792087 0.19280118 0.577073498  0.19823599
#>    15   26 0.247902441 0.663838205 0.19280118 0.577073498  0.32094915
#>    18   46 0.161474274 0.211363552 0.16070859 0.079947471  0.16255572
#>    18   95 0.161474274 0.211363552 0.22777890 0.137536897  0.16653848
#>     2   18 0.089014590 0.231541225 0.16147427 0.211363552  0.13891228
#>    15   48 0.247902441 0.663838205 0.25037057 0.769201342  0.20886294
#>    47   54 0.301606065 0.965221071 0.24998954 0.971132226  0.26457677
#>    46   64 0.160708595 0.079947471 0.27861836 0.005973728  0.23277790
#>    52   66 0.191091307 0.306418229 0.32092273 0.426792087  0.19712997
#>    10    2 0.018484446 0.317315268 0.08901459 0.231541225  0.10508947
#>    84   64 0.335812490 0.000000000 0.27861836 0.005973728  0.22585536
#>    64   25 0.278618362 0.005973728 0.15141678 0.019614574  0.20940503
#>    25   46 0.151416785 0.019614574 0.16070859 0.079947471  0.08108951
#>    75   74 0.142795426 0.723362657 0.11505414 0.702914533  0.17274573
#>    74   85 0.115054138 0.702914533 0.07363540 0.439313252  0.05296297
#>    41   46 0.138228262 0.002837720 0.16070859 0.079947471  0.06702190
#>    75   48 0.142795426 0.723362657 0.25037057 0.769201342  0.19924034
#>    36   47 0.453191439 0.847828033 0.30160606 0.965221071  0.37666462
#>    74   26 0.115054138 0.702914533 0.19280118 0.577073498  0.17200236
#>    85   10 0.073635403 0.439313252 0.01848445 0.317315268  0.10723687
#>     9   10 0.013358076 0.429534162 0.01848445 0.317315268  0.05311724
#>    74    9 0.115054138 0.702914533 0.01335808 0.429534162  0.01941063
#>    23   54 0.173848178 0.793145511 0.24998954 0.971132226  0.15457224
#>    47   86 0.301606065 0.965221071 0.92228447 1.000000000  0.50205082
#>    41   44 0.138228262 0.002837720 0.06167817 0.160083072 -0.04432770
#>    44    2 0.061678165 0.160083072 0.08901459 0.231541225  0.04594080
#>    33    9 0.000000000 0.188512430 0.01335808 0.429534162 -5.04414541
#>    75   23 0.142795426 0.723362657 0.17384818 0.793145511  0.08600487
#>    99   23 0.013596310 0.740626882 0.17384818 0.793145511  0.07091053
#>    33   99 0.000000000 0.188512430 0.01359631 0.740626882 -7.37003457
#>    99   22 0.013596310 0.740626882 0.04613008 0.945400765  0.03624254
#>    22   54 0.046130080 0.945400765 0.24998954 0.971132226 -0.26489706
#>    20   92 0.846059733 0.182398554 0.74571322 0.314643524  0.81825473
#>    78   73 0.959261738 0.505332550 0.98209728 0.378060629  1.15198046
#>    71    5 0.935777388 0.136006320 0.87289992 0.196996766  0.96445800
#>     7   35 0.833275033 0.539804640 0.92268857 0.421914805  0.87071701
#>     5   34 0.872899924 0.196996766 0.88544469 0.112809722  0.87846187
#>    20   51 0.846059733 0.182398554 0.66342646 0.179189620  0.75411009
#>    98   59 0.729980781 0.485163452 0.79681696 0.465596910  0.74139872
#>    98    7 0.729980781 0.485163452 0.83327503 0.539804640  0.77688565
#>    20   13 0.846059733 0.182398554 0.82697024 0.125159586  0.75473857
#>    62  100 0.725001094 0.019192461 0.65452139 0.126461400  0.74059799
#>    71   97 0.935777388 0.136006320 1.00000000 0.023726228  0.94949571
#>    71    1 0.935777388 0.136006320 0.99277685 0.321067002  1.13557017
#>     8   51 0.550912586 0.164104608 0.66342646 0.179189620  0.60855126
#>    55   92 0.530651045 0.344538924 0.74571322 0.314643524  0.64214580
#>    98   29 0.729980781 0.485163452 0.53370977 0.530415197  0.62054867
#>    98   12 0.729980781 0.485163452 0.75731361 0.668558547  0.74790945
#>    98    3 0.729980781 0.485163452 0.63465740 0.570469451  0.72826168
#>    62   13 0.725001094 0.019192461 0.82697024 0.125159586  0.82160171
#>    67   58 0.879380200 0.622385568 0.86639938 0.706390509  0.87952648
#>    55   98 0.530651045 0.344538924 0.72998078 0.485163452  0.64678827
#>    80   77 0.511606632 0.307114877 0.52445198 0.310656582  0.52486613
#>    67   90 0.879380200 0.622385568 0.80973499 0.707398676  0.83721282
#>    80   55 0.511606632 0.307114877 0.53065105 0.344538924  0.51209010
#>   100   56 0.654521390 0.126461400 0.62533596 0.044569267  0.58775747
#>    21   70 0.492097910 0.295143780 0.54877756 0.295675201  0.52105642
#>    61   90 0.746764442 0.717975179 0.80973499 0.707398676  0.77584136
#>    55   21 0.530651045 0.344538924 0.49209791 0.295143780  0.45615904
#>    30    8 0.517435468 0.130233751 0.55091259 0.164104608  0.58161219
#>    30   21 0.517435468 0.130233751 0.49209791 0.295143780  0.47284597
#>    83   58 0.974832334 0.769953479 0.86639938 0.706390509  0.91069840
#>    83   57 0.974832334 0.769953479 0.91049219 0.635537914  0.94088351
#>    73   83 0.982097281 0.378060629 0.97483233 0.769953479  2.31771796
#>    62   69 0.725001094 0.019192461 0.49350707 0.017438755  0.61776956
#>    24   56 0.549833803 0.022247725 0.62533596 0.044569267  0.63485024
#>    24   30 0.549833803 0.022247725 0.51743547 0.130233751  0.57155813
#>    31   12 0.580050078 0.699275190 0.75731361 0.668558547  0.66437204
#>    21   63 0.492097910 0.295143780 0.37245752 0.229915644  0.43385540
#>    55   43 0.530651045 0.344538924 0.37455158 0.398063891  0.47561148
#>    55    4 0.530651045 0.344538924 0.39545113 0.316010645  0.45371116
#>    11   39 0.720893137 0.819939784 0.85286862 0.723629371  0.80221899
#>    87   30 0.435321185 0.059839520 0.51743547 0.130233751  0.49212577
#>     4   72 0.395451126 0.316010645 0.36815604 0.358331031  0.43062266
#>    29   93 0.533709767 0.530415197 0.50377114 0.701475422  0.53958047
#>    79    4 0.348628312 0.281157443 0.39545113 0.316010645  0.39288674
#>    17   63 0.379820924 0.102030715 0.37245752 0.229915644  0.39154149
#>    21   53 0.492097910 0.295143780 0.43748185 0.119640392  0.46183801
#>    82   40 0.330717817 0.368899455 0.35155022 0.375936763  0.34701086
#>    82   72 0.330717817 0.368899455 0.36815604 0.358331031  0.34285365
#>    40   43 0.351550216 0.375936763 0.37455158 0.398063891  0.37154948
#>    40   66 0.351550216 0.375936763 0.32092273 0.426792087  0.34446531
#>    17   53 0.379820924 0.102030715 0.43748185 0.119640392  0.41485625
#>    42   88 0.895959051 0.954616364 0.92665936 0.842330696  0.90699425
#>    94   83 0.977659049 0.895067332 0.97483233 0.769953479  0.98989931
#>    94    1 0.977659049 0.895067332 0.99277685 0.321067002  5.56803109
#>    60   17 0.359802618 0.062620612 0.37982092 0.102030715  0.39781495
#>    79   89 0.348628312 0.281157443 0.31329101 0.262481584  0.30642541
#>    82   28 0.330717817 0.368899455 0.36753678 0.321418848  0.31727010
#>    29   96 0.533709767 0.530415197 0.41730844 0.644903039  0.50012600
#>    37   93 0.387718645 0.810344984 0.50377114 0.701475422  0.42257550
#>    31   32 0.580050078 0.699275190 0.54274474 0.784879948  0.63719202
#>    76   50 0.602977744 0.864454291 0.85832283 0.934771631  0.71080962
#>    89   63 0.313291011 0.262481584 0.37245752 0.229915644  0.32710626
#>    52   89 0.191091307 0.306418229 0.31329101 0.262481584  0.26512935
#>    60   87 0.359802618 0.062620612 0.43532119 0.059839520  0.39603680
#>    60   49 0.359802618 0.062620612 0.30970296 0.073119839  0.34086367
#>    32   36 0.542744739 0.784879948 0.45319144 0.847828033  0.52779143
#>    76   11 0.602977744 0.864454291 0.72089314 0.819939784  0.63699204
#>    50   86 0.858322825 0.934771631 0.92228447 1.000000000  0.83827558
#>    42   45 0.895959051 0.954616364 0.87396174 0.849163641  0.90153337
#>    89   91 0.313291011 0.262481584 0.29480341 0.183402073  0.30098383
#>    52   82 0.191091307 0.306418229 0.33071782 0.368899455  0.26668104
#>    60   84 0.359802618 0.062620612 0.33581249 0.000000000  0.34162197
#>    60   81 0.359802618 0.062620612 0.31944751 0.059687468  0.33800794
#>    91   14 0.294803409 0.183402073 0.22671998 0.213306947  0.26448139
#>    84   62 0.335812490 0.000000000 0.72500109 0.019192461  0.67486171
#>    60   68 0.359802618 0.062620612 0.45177023 0.013583659  0.39294485
#>    64   49 0.278618362 0.005973728 0.30970296 0.073119839  0.28402338
#>    86   94 0.922284468 1.000000000 0.97765905 0.895067332  0.95545922
#>    37   36 0.387718645 0.810344984 0.45319144 0.847828033  0.45605586
#>    16   19 0.223483614 0.174605602 0.21818892 0.220518257  0.20450538
#>    95   91 0.227778901 0.137536897 0.29480341 0.183402073  0.28493290
#>    52   27 0.191091307 0.306418229 0.26635222 0.245091357  0.22357799
#>    95   16 0.227778901 0.137536897 0.22348361 0.174605602  0.26145993
#>    96   48 0.417308439 0.644903039 0.25037057 0.769201342  0.34095970
#>    64   95 0.278618362 0.005973728 0.22777890 0.137536897  0.23760396
#>    64   81 0.278618362 0.005973728 0.31944751 0.059687468  0.30950139
#>    19   18 0.218188919 0.220518257 0.16147427 0.211363552  0.18324911
#>    85   26 0.073635403 0.439313252 0.19280118 0.577073498  0.19823599
#>    95   46 0.227778901 0.137536897 0.16070859 0.079947471  0.16255572
#>    96   15 0.417308439 0.644903039 0.24790244 0.663838205  0.33931598
#>    48   54 0.250370567 0.769201342 0.24998954 0.971132226  0.26457677
#>    44   18 0.061678165 0.160083072 0.16147427 0.211363552  0.13207830
#>    85   66 0.073635403 0.439313252 0.32092273 0.426792087  0.19712997
#>    52    2 0.191091307 0.306418229 0.08901459 0.231541225  0.10508947
#>    41   64 0.138228262 0.002837720 0.27861836 0.005973728  0.22585536
#>    41   25 0.138228262 0.002837720 0.15141678 0.019614574  0.20940503
#>    15   74 0.247902441 0.663838205 0.11505414 0.702914533  0.17274573
#>    44   46 0.061678165 0.160083072 0.16070859 0.079947471  0.06702190
#>    33   10 0.000000000 0.188512430 0.01848445 0.317315268  0.02485951
#>    23   48 0.173848178 0.793145511 0.25037057 0.769201342  0.19924034
#>    37   47 0.387718645 0.810344984 0.30160606 0.965221071  0.37666462
#>    99   75 0.013596310 0.740626882 0.14279543 0.723362657  0.08384795
#>    47   76 0.301606065 0.965221071 0.60297774 0.864454291  0.50430287
#>    52   10 0.191091307 0.306418229 0.01848445 0.317315268  0.10723687
#>    85    9 0.073635403 0.439313252 0.01335808 0.429534162  0.01941063
#>    54   86 0.249989542 0.971132226 0.92228447 1.000000000  0.50205082
#>    33   44 0.000000000 0.188512430 0.06167817 0.160083072 -0.04432770
#>    33    2 0.000000000 0.188512430 0.08901459 0.231541225  0.04594080
#>    99    9 0.013596310 0.740626882 0.01335808 0.429534162 -5.04414541
#>    99   74 0.013596310 0.740626882 0.11505414 0.702914533  0.01351683
#>    22   23 0.046130080 0.945400765 0.17384818 0.793145511  0.07091053
#>    65   99 0.008272485 0.942035772 0.01359631 0.740626882 -7.37003457
#>    65   22 0.008272485 0.942035772 0.04613008 0.945400765  0.03624254
#>    65   54 0.008272485 0.942035772 0.24998954 0.971132226 -0.26489706
#>     1   73 0.992776850 0.321067002 0.98209728 0.378060629  0.90357064
#>    38   35 0.959240659 0.443250718 0.92268857 0.421914805  0.95663575
#>    20    5 0.846059733 0.182398554 0.87289992 0.196996766  0.81825473
#>     7   67 0.833275033 0.539804640 0.87938020 0.622385568  0.80486394
#>    57   67 0.910492190 0.635537914 0.87938020 0.622385568  0.92175549
#>    71    6 0.935777388 0.136006320 0.91682226 0.009644253  0.94949571
#>     8  100 0.550912586 0.164104608 0.65452139 0.126461400  0.60855126
#>     3   31 0.634657395 0.570469451 0.58005008 0.699275190  0.55884346
#>    62   24 0.725001094 0.019192461 0.54983380 0.022247725  0.61776956
#>    11   90 0.720893137 0.819939784 0.80973499 0.707398676  0.80221899
#>    11   31 0.720893137 0.819939784 0.58005008 0.699275190  0.65877593
#>    79   82 0.348628312 0.281157443 0.33071782 0.368899455  0.30642541
#>    37   96 0.387718645 0.810344984 0.41730844 0.644903039  0.42257550
#>    32   76 0.542744739 0.784879948 0.60297774 0.864454291  0.52779143
#>    50   42 0.858322825 0.934771631 0.89595905 0.954616364  0.83827558
#>    84    6 0.335812490 0.000000000 0.91682226 0.009644253  0.67486171
#>    52   19 0.191091307 0.306418229 0.21818892 0.220518257  0.22357799
#>    75   15 0.142795426 0.723362657 0.24790244 0.663838205  0.20886294
#>    48   47 0.250370567 0.769201342 0.30160606 0.965221071  0.26457677
#>    85   52 0.073635403 0.439313252 0.19109131 0.306418229  0.19712997
#>    41   84 0.138228262 0.002837720 0.33581249 0.000000000  0.22585536
#>    33   41 0.000000000 0.188512430 0.13822826 0.002837720 -0.04432770
#>    65   33 0.008272485 0.942035772 0.00000000 0.188512430 -7.37003457
#>          my1         mx2          my2 bp1 bp2
#>   0.33384877  0.95663575   0.40573542   0   0
#>   0.31437860  0.90357064   0.33384877   0   0
#>   0.31659873  0.87555563   0.31437860   0   0
#>   0.26549390  0.86552631   0.31659873   0   0
#>   0.36895201  0.86552631   0.31659873   0   0
#>   0.48743890  0.91661017   0.47430611   0   0
#>   0.47430611  1.15198046   0.47422620   0   0
#>   0.47422620  0.95663575   0.40573542   0   0
#>   0.58909166  1.09431568   0.63015438   0   0
#>   0.22848095  0.87555563   0.31437860   0   0
#>   0.15732726  0.93498411   0.07152259   0   0
#>   0.47534974  0.83379306   0.36895201   0   0
#>   0.15479738  0.86177673   0.14535415   0   0
#>   0.14535415  0.84008763   0.04266002   0   0
#>   0.21682142  0.64964121   0.28028511   0   0
#>   0.40023117  0.83379306   0.36895201   0   0
#>   0.52144886  0.87071701   0.47534974   0   0
#>   0.18105193  0.74084537   0.13899854   0   0
#>   0.28028511  0.59925328   0.23069166   0   0
#>   0.60982742  0.81722191   0.64249794   0   0
#>   0.55399528  0.88665631   0.48743890   0   0
#>   0.56552047  0.90486726   0.55399528   0   0
#>   0.10622858  0.74084537   0.13899854   0   0
#>   0.06934575  2.92817964 -11.61811030   0   1
#>   0.04266002  0.93498411   0.07152259   0   0
#>   0.17577769  4.49673121   0.25742868   0   1
#>   0.16134121  0.74084537   0.13899854   0   0
#>   0.35810511  0.64964121   0.28028511   0   0
#>   0.45879240  0.55884346   0.61430634   0   0
#>   0.57622576  0.80486394   0.60982742   0   0
#>   0.57915403  0.66437204   0.65904528   0   0
#>   0.02828100  0.84008763   0.04266002   0   0
#>   0.66541357  0.94088351   0.70359713   0   0
#>   0.39150226  0.74139872   0.40023117   0   0
#>   0.64249794  0.77584136   0.69834804   0   0
#>   0.28408938  0.54922003   0.32363331   0   0
#>   0.32363331  0.64214580   0.35810511   0   0
#>   0.65887507  0.83790817   0.69795737   0   0
#>   0.06953900  0.63485024  -0.12646560   0   0
#>   0.27046028  0.52486613   0.28408938   0   0
#>   0.33042655  0.54922003   0.32363331   0   0
#>   0.35130420  0.52067165   0.27046028   0   0
#>   0.10410856  0.58161219   0.10028247   0   0
#>   0.22942274  0.59925328   0.23069166   0   0
#>   0.43610896  0.47561148   0.43840794   0   0
#>   0.67514789  0.77584136   0.69834804   0   0
#>   0.79625569  0.89199234   0.78159478   0   0
#>   0.36293725  0.43385540   0.25963594   0   0
#>   0.61430634  0.53958047   0.61959269   0   0
#>   0.20778432  0.52105642   0.22942274   0   0
#>   0.77456422  0.91069840   0.75509032   0   0
#>   0.75509032  0.83790817   0.69795737   0   0
#>   0.63015438  0.92175549   0.56552047   0   0
#>   0.59883426  5.56803109   0.72876755   0   0
#>  -1.10575132  0.51727449   0.07133231   0   0
#>   0.08761867  0.58161219   0.10028247   0   0
#>   0.78159478  0.90330185   0.77456422   0   0
#>   0.20682602  0.48234198   0.08808008   0   0
#>   0.25963594  0.39288674   0.27057754   0   0
#>   0.43840794  0.40831180   0.51933872   0   0
#>   0.37453814  0.43062266   0.36865734   0   0
#>   0.79280260  0.83790817   0.69795737   0   0
#>   0.78323480  0.65877593   0.74991445   0   0
#>  -2.00935793  0.62148302  -1.59593871   0   0
#>   0.07666738  0.51727449   0.07133231   0   0
#>   0.74991445  0.66716234   0.67514789   0   0
#>   0.36865734  0.37154948   0.37816596   0   0
#>   0.61959269  0.54288074   0.73400827   0   0
#>   0.33957817  0.34285365   0.34029423   0   0
#>   0.16686002  0.46183801   0.20831070   0   0
#>   0.51933872  0.40528823   0.51986246   0   0
#>   0.73400827  0.46550206   0.77016608   0   0
#>   0.35502100  0.37154948   0.37816596   0   0
#>   0.40632022  0.40528823   0.51986246   0   0
#>  -1.59593871  0.46981750   0.04605410   0   0
#>   0.04605410  0.49212577   0.07666738   0   0
#>   0.16496360  0.32662866   0.13155418   0   0
#>   0.09051841  0.48234198   0.08808008   0   0
#>   0.29266355  0.38553609   0.33957817   0   0
#>   0.90061812  0.98989931   0.83220193   0   0
#>   0.89729377  0.89199234   0.78159478   0   0
#>   0.83220193  0.90330185   0.77456422   0   0
#>   0.06810149  0.41485625   0.09051841   0   0
#>   0.31824171  0.26668104   0.32475018   0   0
#>   0.32045540  0.37644672   0.29266355   0   0
#>   0.61268744  0.42257550   0.73121215   0   0
#>   0.39892195  0.24043624   0.38339930   0   0
#>   0.77510779  0.63699204   0.77612395   0   0
#>   0.97166085  0.78915085   0.87790280   0   0
#>   0.51986246  0.32169972   0.55680117   0   0
#>   0.21755097  0.35860526   0.16496360   0   0
#>   0.32043453  0.30098383   0.22365800   0   0
#>   0.87790280  0.80473892   0.79625569   0   0
#>   0.01981695  0.46981750   0.04605410   0   0
#>   0.09702981  0.32662866   0.13155418   0   0
#>   0.85878226  0.50430287   1.07039123   0   0
#>   1.01840351  0.95545922   0.95042948   0   0
#>   0.89843290  0.78915085   0.87790280   0   0
#>   0.24770579  0.39288674   0.27057754   0   0
#>   0.22365800  0.26448139   0.20682302   0   0
#>   0.03368002  0.30950139   0.02487327   0   0
#>   0.08340297  0.28402338   0.04423973   0   0
#>   0.13155418  0.28493290   0.12592092   0   0
#>   0.20682302  0.23986826   0.23751334   0   0
#>  -2.91968983  0.82160171   0.02828100   0   0
#>   0.01401808  0.62996957  -2.00935793   0   0
#>   1.20692273  0.59228253   1.33351942   0   0
#>   0.95042948  0.91915300   0.90061812   0   0
#>   0.76690146  0.46550206   0.77016608   0   0
#>   0.19567863  0.23986826   0.23751334   0   0
#>   0.26944230  0.23986826   0.23751334   0   0
#>   0.16022285  0.25763485   0.19123572   0   0
#>   0.71661492  0.29753449   0.86157948   0   0
#>   0.19123572  0.20450538   0.19567863   0   0
#>   0.06572912  0.28493290   0.12592092   0   0
#>   0.19438765  0.16653848   0.14922395   0   0
#>   0.25672033  0.13891228   0.27053475   0   0
#>   0.55680117  0.32094915   0.55656996   0   0
#>   0.45195157  0.32094915   0.55656996   0   0
#>   0.55656996  0.17200236   0.65116092   0   0
#>   0.14564698  0.13207830   0.14582455   0   0
#>   0.14922395  0.16255572   0.14564698   0   0
#>   0.27053475  0.11398803   0.18102973   0   0
#>   0.71746318  0.33931598   0.71440732   0   0
#>   0.87019395  0.50205082   2.94382995   0   0
#>   0.06386420  0.21796167   0.04024802   0   0
#>   0.43010807  0.24043624   0.38339930   0   0
#>   0.31664398  0.02485951   0.25067262   0   0
#>  -0.77597695  0.30950139   0.02487327   0   0
#>  -0.03954308  0.21796167   0.04024802   0   0
#>   0.06132757  0.21796167   0.04024802   0   0
#>   0.65368818  0.08384795   0.77429284   0   0
#>   0.57761607  0.01941063   0.58288803   0   0
#>   0.06542879  0.08108951   0.06132757   0   0
#>   0.74004570  0.20886294   0.71746318   0   0
#>   0.90557659  0.50430287   1.07039123   0   0
#>   0.65116092  0.05296297   0.57761607   0   0
#>   0.35065833  0.05311724   0.37512390   0   0
#>   0.37512390 -0.62612329   0.34409491   0   0
#>   0.58288803  0.01351683   0.58508049   0   0
#>   0.90667132  0.23994972   0.87014748   0   0
#>   2.94382995  0.59228253   1.33351942   0   0
#>   0.01122164  0.06702190   0.06542879   0   0
#>   0.20706129  0.11398803   0.18102973   0   0
#>   0.58895364 -0.62612329   0.34409491   0   0
#>   0.79043448  0.19924034   0.74004570   0   0
#>   0.83649236  0.08600487   0.79043448   0   0
#>   0.64623074 -5.04414541   0.58895364   0   0
#>   0.84200030  0.07091053   0.83649236   0   0
#>   4.22994882  0.15457224   0.90667132   0   0
#>   0.26549390  0.75411009   0.21682142   0   0
#>   0.47422620  1.79217313   0.58909166   0   0
#>   0.22848095  0.89543970   0.15732726   0   0
#>   0.47534974  0.88665631   0.48743890   0   0
#>   0.15479738  0.89543970   0.15732726   0   0
#>   0.21682142  0.75473857   0.18105193   0   0
#>   0.40023117  0.77688565   0.52144886   0   0
#>   0.52144886  0.74790945   0.57622576   0   0
#>   0.18105193  0.86177673   0.14535415   0   0
#>   0.10622858  0.68475707   0.06953900   0   0
#>   0.06934575  1.13557017   0.17577769   0   0
#>   0.17577769  0.96445800   0.22848095   0   0
#>   0.16134121  0.59925328   0.23069166   0   0
#>   0.35810511  0.64678827   0.39150226   0   0
#>   0.45879240  0.61531883   0.43610896   0   0
#>   0.57622576  0.72826168   0.57915403   0   0
#>   0.57915403  0.62054867   0.45879240   0   0
#>   0.02828100  0.74059799   0.10622858   0   0
#>   0.66541357  0.83721282   0.65887507   0   0
#>   0.39150226  0.61531883   0.43610896   0   0
#>   0.28408938  0.51209010   0.33042655   0   0
#>   0.65887507  0.81722191   0.64249794   0   0
#>   0.33042655  0.47106357   0.35130420   0   0
#>   0.10410856  0.68475707   0.06953900   0   0
#>   0.22942274  0.52067165   0.27046028   0   0
#>   0.69834804  0.79009890   0.78323480   0   0
#>   0.36293725  0.47106357   0.35130420   0   0
#>   0.10028247  0.47284597   0.20778432   0   0
#>   0.20778432  0.46660888   0.20682602   0   0
#>   0.75509032  0.94088351   0.70359713   0   0
#>   0.70359713  1.09431568   0.63015438   0   0
#>   0.59883426  1.79217313   0.58909166   0   0
#>  -1.10575132  0.62148302  -1.59593871   0   0
#>  -0.12646560  0.57155813   0.08761867   0   0
#>   0.08761867  0.51727449   0.07133231   0   0
#>   0.65904528  0.66716234   0.67514789   0   0
#>   0.25963594  0.46183801   0.20831070   0   0
#>   0.43840794  0.45371116   0.37453814   0   0
#>   0.37453814  0.45615904   0.36293725   0   0
#>   0.79280260  0.80473892   0.79625569   0   0
#>   0.07666738  0.48234198   0.08808008   0   0
#>   0.36865734  0.38553609   0.33957817   0   0
#>   0.61959269  0.50012600   0.61268744   0   0
#>   0.27057754  0.37644672   0.29266355   0   0
#>   0.16686002  0.35860526   0.16496360   0   0
#>   0.20831070  0.46660888   0.20682602   0   0
#>   0.35502100  0.33218086   0.39892195   0   0
#>   0.34029423  0.34701086   0.35502100   0   0
#>   0.37816596  0.34446531   0.40632022   0   0
#>   0.40632022  0.33218086   0.39892195   0   0
#>   0.09051841  0.39154149   0.16686002   0   0
#>   0.89729377  0.91915300   0.90061812   0   0
#>   0.83220193  5.56803109   0.72876755   0   0
#>   0.72876755  7.30898272   0.77462007   0   1
#>   0.06810149  0.34086367   0.09702981   0   0
#>   0.31824171  0.34370383   0.24770579   0   0
#>   0.32045540  0.34285365   0.34029423   0   0
#>   0.61268744  0.40831180   0.51933872   0   0
#>   0.73121215  0.45605586   0.76690146   0   0
#>   0.77510779  0.54288074   0.73400827   0   0
#>   0.97166085  0.64602283   1.20692273   0   0
#>   0.21755097  0.34370383   0.24770579   0   0
#>   0.32043453  0.26668104   0.32475018   0   0
#>   0.01981695  0.39781495   0.06810149   0   0
#>   0.09702981  0.33800794   0.08340297   0   0
#>   0.85878226  0.46550206   0.77016608   0   0
#>   0.77612395  0.71080962   0.97166085   0   0
#>   1.01840351  0.64602283   1.20692273   0   0
#>   0.89843290  0.90699425   0.89729377   0   0
#>   0.22365800  0.32710626   0.21755097   0   0
#>   0.32475018  0.24043624   0.38339930   0   0
#>   0.03368002  0.39294485   0.01401808   0   0
#>   0.08340297  0.34162197   0.03368002   0   0
#>   0.20682302  0.25763485   0.19123572   0   0
#>  -2.91968983  0.62996957  -2.00935793   0   0
#>   0.01401808  0.39603680   0.01981695   0   0
#>   0.04423973  0.23760396   0.06572912   0   0
#>   0.95042948  8.40954104   4.88406302   0   1
#>   0.76690146  0.37666462   0.90557659   0   0
#>   0.19567863  0.19331066   0.19438765   0   0
#>   0.12592092  0.26145993   0.16022285   0   0
#>   0.26944230  0.26512935   0.32043453   0   0
#>   0.16022285  0.16653848   0.14922395   0   0
#>   0.71661492  0.33931598   0.71440732   0   0
#>   0.06572912  0.23277790   0.06386420   0   0
#>   0.02487327  0.28402338   0.04423973   0   0
#>   0.25672033  0.19331066   0.19438765   0   0
#>   0.45195157  0.05296297   0.57761607   0   0
#>   0.14564698  0.23277790   0.06386420   0   0
#>   0.71440732  0.32169972   0.55680117   0   0
#>   0.87019395  0.23994972   0.87014748   0   0
#>   0.14582455  0.11398803   0.18102973   0   0
#>   0.43010807  0.19823599   0.45195157   0   0
#>   0.31664398  0.13891228   0.27053475   0   0
#>  -0.77597695  0.20940503  -0.03954308   0   0
#>  -0.03954308  0.08108951   0.06132757   0   0
#>   0.65368818  0.17200236   0.65116092   0   0
#>   0.06542879  0.13207830   0.14582455   0   0
#>   0.25067262 -0.62612329   0.34409491   0   0
#>   0.74004570  0.23994972   0.87014748   0   0
#>   0.90557659  0.29753449   0.86157948   0   0
#>   0.77429284  0.08600487   0.79043448   0   0
#>   1.07039123  0.59228253   1.33351942   0   0
#>   0.35065833  0.10508947   0.31664398   0   0
#>   0.58288803  0.05311724   0.37512390   0   0
#>   2.94382995  0.43829884   4.42853493   0   1
#>   0.01122164  0.04594080   0.20706129   0   0
#>   0.20706129  0.02485951   0.25067262   0   0
#>   0.58895364  0.01351683   0.58508049   0   0
#>   0.58508049  0.08384795   0.77429284   0   0
#>   0.83649236  0.15457224   0.90667132   0   0
#>   0.64623074  0.03624254   0.84200030   0   0
#>   0.84200030 -0.26489706   4.22994882   0   0
#>   4.22994882 -0.75577977   8.30792769   0   1
#>   0.33384877  2.31771796   0.59883426   0   0
#>   0.40573542  0.91661017   0.47430611   0   0
#>   0.26549390  0.87846187   0.15479738   0   0
#>   0.60982742  0.90486726   0.55399528   0   0
#>   0.56552047  0.87952648   0.66541357   0   0
#>   0.06934575  0.93498411   0.07152259   0   0
#>   0.16134121  0.58775747   0.10410856   0   0
#>   0.61430634  0.66437204   0.65904528   0   0
#>  -1.10575132  0.63485024  -0.12646560   0   0
#>   0.79280260  0.79009890   0.78323480   0   0
#>   0.74991445  0.63719202   0.77510779   0   0
#>   0.31824171  0.31727010   0.32045540   0   0
#>   0.73121215  0.34095970   0.71661492   0   0
#>   0.85878226  0.63699204   0.77612395   0   0
#>   1.01840351  0.90153337   0.89843290   0   0
#>  -2.91968983  0.70342328  -4.64035719   0   1
#>   0.26944230  0.18324911   0.25672033   0   0
#>   0.71746318  0.17274573   0.65368818   0   0
#>   0.87019395  0.29753449   0.86157948   0   0
#>   0.43010807  0.10723687   0.35065833   0   0
#>  -0.77597695  0.15318198  -5.83606593   0   1
#>   0.01122164 -3.50956208  -2.56852294   0   1
#>   0.64623074 -8.69697354   0.66079841   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.992776850 0.321067002
#>   [2,] 0.089014590 0.231541225
#>   [3,] 0.634657395 0.570469451
#>   [4,] 0.395451126 0.316010645
#>   [5,] 0.872899924 0.196996766
#>   [6,] 0.916822256 0.009644253
#>   [7,] 0.833275033 0.539804640
#>   [8,] 0.550912586 0.164104608
#>   [9,] 0.013358076 0.429534162
#>  [10,] 0.018484446 0.317315268
#>  [11,] 0.720893137 0.819939784
#>  [12,] 0.757313613 0.668558547
#>  [13,] 0.826970240 0.125159586
#>  [14,] 0.226719984 0.213306947
#>  [15,] 0.247902441 0.663838205
#>  [16,] 0.223483614 0.174605602
#>  [17,] 0.379820924 0.102030715
#>  [18,] 0.161474274 0.211363552
#>  [19,] 0.218188919 0.220518257
#>  [20,] 0.846059733 0.182398554
#>  [21,] 0.492097910 0.295143780
#>  [22,] 0.046130080 0.945400765
#>  [23,] 0.173848178 0.793145511
#>  [24,] 0.549833803 0.022247725
#>  [25,] 0.151416785 0.019614574
#>  [26,] 0.192801179 0.577073498
#>  [27,] 0.266352224 0.245091357
#>  [28,] 0.367536778 0.321418848
#>  [29,] 0.533709767 0.530415197
#>  [30,] 0.517435468 0.130233751
#>  [31,] 0.580050078 0.699275190
#>  [32,] 0.542744739 0.784879948
#>  [33,] 0.000000000 0.188512430
#>  [34,] 0.885444693 0.112809722
#>  [35,] 0.922688570 0.421914805
#>  [36,] 0.453191439 0.847828033
#>  [37,] 0.387718645 0.810344984
#>  [38,] 0.959240659 0.443250718
#>  [39,] 0.852868618 0.723629371
#>  [40,] 0.351550216 0.375936763
#>  [41,] 0.138228262 0.002837720
#>  [42,] 0.895959051 0.954616364
#>  [43,] 0.374551580 0.398063891
#>  [44,] 0.061678165 0.160083072
#>  [45,] 0.873961741 0.849163641
#>  [46,] 0.160708595 0.079947471
#>  [47,] 0.301606065 0.965221071
#>  [48,] 0.250370567 0.769201342
#>  [49,] 0.309702959 0.073119839
#>  [50,] 0.858322825 0.934771631
#>  [51,] 0.663426465 0.179189620
#>  [52,] 0.191091307 0.306418229
#>  [53,] 0.437481850 0.119640392
#>  [54,] 0.249989542 0.971132226
#>  [55,] 0.530651045 0.344538924
#>  [56,] 0.625335964 0.044569267
#>  [57,] 0.910492190 0.635537914
#>  [58,] 0.866399379 0.706390509
#>  [59,] 0.796816961 0.465596910
#>  [60,] 0.359802618 0.062620612
#>  [61,] 0.746764442 0.717975179
#>  [62,] 0.725001094 0.019192461
#>  [63,] 0.372457518 0.229915644
#>  [64,] 0.278618362 0.005973728
#>  [65,] 0.008272485 0.942035772
#>  [66,] 0.320922725 0.426792087
#>  [67,] 0.879380200 0.622385568
#>  [68,] 0.451770231 0.013583659
#>  [69,] 0.493507066 0.017438755
#>  [70,] 0.548777559 0.295675201
#>  [71,] 0.935777388 0.136006320
#>  [72,] 0.368156039 0.358331031
#>  [73,] 0.982097281 0.378060629
#>  [74,] 0.115054138 0.702914533
#>  [75,] 0.142795426 0.723362657
#>  [76,] 0.602977744 0.864454291
#>  [77,] 0.524451981 0.310656582
#>  [78,] 0.959261738 0.505332550
#>  [79,] 0.348628312 0.281157443
#>  [80,] 0.511606632 0.307114877
#>  [81,] 0.319447507 0.059687468
#>  [82,] 0.330717817 0.368899455
#>  [83,] 0.974832334 0.769953479
#>  [84,] 0.335812490 0.000000000
#>  [85,] 0.073635403 0.439313252
#>  [86,] 0.922284468 1.000000000
#>  [87,] 0.435321185 0.059839520
#>  [88,] 0.926659364 0.842330696
#>  [89,] 0.313291011 0.262481584
#>  [90,] 0.809734986 0.707398676
#>  [91,] 0.294803409 0.183402073
#>  [92,] 0.745713221 0.314643524
#>  [93,] 0.503771139 0.701475422
#>  [94,] 0.977659049 0.895067332
#>  [95,] 0.227778901 0.137536897
#>  [96,] 0.417308439 0.644903039
#>  [97,] 1.000000000 0.023726228
#>  [98,] 0.729980781 0.485163452
#>  [99,] 0.013596310 0.740626882
#> [100,] 0.654521390 0.126461400
#> 
#> $tri.obj
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,73,35), [2,3,62]
#> 2: (38,35,73), [1,9,8]
#> 3: (5,1,35), [1,4,11]
#> 4: (92,5,35), [3,6,5]
#> 5: (20,5,92), [4,16,14]
#> 6: (59,92,35), [4,13,17]
#> 7: (7,35,78), [8,22,13]
#> 8: (35,38,78), [9,7,2]
#> 9: (78,38,73), [2,10,8]
#> 10: (73,83,78), [61,9,62]
#> 11: (71,1,5), [3,12,29]
#> 12: (5,34,71), [27,11,14]
#> 13: (7,59,35), [6,7,18]
#> 14: (5,20,34), [15,12,5]
#> 15: (20,13,34), [28,14,19]
#> 16: (20,92,51), [20,19,5]
#> 17: (98,92,59), [6,18,37]
#> 18: (98,59,7), [13,33,17]
#> 19: (20,51,13), [25,15,16]
#> 20: (92,70,51), [39,16,31]
#> 21: (7,67,12), [38,33,22]
#> 22: (67,7,78), [7,23,21]
#> 23: (57,67,78), [22,61,36]
#> 24: (62,13,100), [25,43,35]
#> 25: (100,13,51), [19,30,24]
#> 26: (71,6,97), [0,29,27]
#> 27: (6,71,34), [12,28,26]
#> 28: (13,6,34), [27,15,35]
#> 29: (71,97,1), [0,11,26]
#> 30: (8,100,51), [25,39,48]
#> 31: (55,70,92), [20,37,41]
#> 32: (98,3,29), [55,50,34]
#> 33: (98,7,12), [21,34,18]
#> 34: (98,12,3), [66,32,33]
#> 35: (62,6,13), [28,24,125]
#> 36: (67,57,58), [60,42,23]
#> 37: (55,92,98), [17,50,31]
#> 38: (67,90,12), [52,21,42]
#> 39: (8,51,70), [20,49,30]
#> 40: (80,70,77), [41,45,44]
#> 41: (77,70,55), [31,45,40]
#> 42: (67,58,90), [47,38,36]
#> 43: (100,56,62), [64,24,48]
#> 44: (21,70,80), [40,46,49]
#> 45: (80,77,55), [41,46,40]
#> 46: (55,21,80), [44,45,54]
#> 47: (39,90,58), [42,59,72]
#> 48: (100,8,56), [56,43,30]
#> 49: (21,8,70), [39,44,57]
#> 50: (98,29,55), [70,37,32]
#> 51: (31,12,61), [52,77,66]
#> 52: (61,12,90), [38,74,51]
#> 53: (11,39,45), [67,111,72]
#> 54: (55,4,21), [69,46,71]
#> 55: (3,31,29), [79,32,66]
#> 56: (30,56,8), [48,57,65]
#> 57: (30,8,21), [49,68,56]
#> 58: (88,39,83), [59,98,67]
#> 59: (83,39,58), [47,60,58]
#> 60: (83,58,57), [36,61,59]
#> 61: (83,57,78), [23,10,60]
#> 62: (73,1,83), [99,10,1]
#> 63: (62,24,69), [73,91,64]
#> 64: (24,62,56), [43,65,63]
#> 65: (24,56,30), [56,73,64]
#> 66: (31,3,12), [34,51,55]
#> 67: (45,39,88), [58,97,53]
#> 68: (21,53,30), [84,57,83]
#> 69: (21,4,63), [80,83,54]
#> 70: (55,29,43), [85,71,50]
#> 71: (55,43,4), [78,54,70]
#> 72: (11,90,39), [47,53,74]
#> 73: (30,69,24), [63,65,76]
#> 74: (90,11,61), [77,52,72]
#> 75: (84,62,68), [91,126,125]
#> 76: (87,69,30), [73,84,92]
#> 77: (11,31,61), [51,74,106]
#> 78: (4,43,72), [89,81,71]
#> 79: (29,31,93), [86,103,55]
#> 80: (79,63,4), [69,95,118]
#> 81: (4,72,28), [88,95,78]
#> 82: (17,53,63), [83,93,94]
#> 83: (21,63,53), [82,68,69]
#> 84: (87,30,53), [68,94,76]
#> 85: (29,96,43), [108,70,103]
#> 86: (31,32,93), [131,79,106]
#> 87: (82,72,40), [89,104,88]
#> 88: (82,28,72), [81,87,102]
#> 89: (40,72,43), [78,90,87]
#> 90: (40,43,66), [108,104,89]
#> 91: (62,69,68), [92,75,63]
#> 92: (68,69,87), [76,112,91]
#> 93: (63,91,17), [123,82,109]
#> 94: (17,87,53), [84,82,100]
#> 95: (79,4,28), [81,102,80]
#> 96: (42,88,94), [98,129,97]
#> 97: (42,45,88), [67,96,117]
#> 98: (94,88,83), [58,99,96]
#> 99: (94,83,1), [62,0,98]
#> 100: (60,87,17), [94,113,112]
#> 101: (79,82,89), [120,118,102]
#> 102: (82,79,28), [95,88,101]
#> 103: (29,93,96), [105,85,79]
#> 104: (40,66,82), [140,87,90]
#> 105: (37,96,93), [103,130,138]
#> 106: (31,11,32), [115,86,77]
#> 107: (76,11,50), [111,128,115]
#> 108: (43,96,66), [145,90,85]
#> 109: (89,91,63), [93,118,119]
#> 110: (52,27,89), [119,120,136]
#> 111: (50,11,45), [53,117,107]
#> 112: (60,68,87), [92,100,126]
#> 113: (60,17,49), [123,122,100]
#> 114: (32,76,36), [172,131,115]
#> 115: (76,32,11), [106,107,114]
#> 116: (50,42,86), [129,128,117]
#> 117: (42,50,45), [111,97,116]
#> 118: (89,63,79), [80,101,109]
#> 119: (89,27,91), [124,109,110]
#> 120: (52,89,82), [101,140,110]
#> 121: (60,81,84), [142,126,122]
#> 122: (60,49,81), [127,121,113]
#> 123: (17,91,49), [134,113,93]
#> 124: (91,27,14), [135,139,119]
#> 125: (84,6,62), [35,75,0]
#> 126: (60,84,68), [75,112,121]
#> 127: (64,81,49), [122,141,142]
#> 128: (50,86,76), [132,107,116]
#> 129: (86,42,94), [96,0,116]
#> 130: (37,93,36), [131,168,105]
#> 131: (36,93,32), [86,114,130]
#> 132: (47,76,86), [128,178,172]
#> 133: (16,14,19), [135,143,139]
#> 134: (95,49,91), [123,137,141]
#> 135: (19,14,27), [124,136,133]
#> 136: (52,19,27), [135,110,144]
#> 137: (95,91,16), [139,149,134]
#> 138: (96,37,48), [169,152,105]
#> 139: (91,14,16), [133,137,124]
#> 140: (52,82,66), [104,158,120]
#> 141: (64,49,95), [134,154,127]
#> 142: (64,84,81), [121,127,160]
#> 143: (19,18,16), [149,133,144]
#> 144: (19,52,18), [150,143,136]
#> 145: (96,15,66), [147,108,152]
#> 146: (85,66,26), [147,164,158]
#> 147: (66,15,26), [170,146,145]
#> 148: (95,18,46), [157,154,149]
#> 149: (16,18,95), [148,137,143]
#> 150: (52,2,18), [156,144,159]
#> 151: (75,15,48), [152,167,163]
#> 152: (96,48,15), [151,145,138]
#> 153: (48,47,54), [178,177,169]
#> 154: (95,46,64), [155,141,148]
#> 155: (25,64,46), [154,162,161]
#> 156: (44,18,2), [150,180,157]
#> 157: (44,46,18), [148,156,165]
#> 158: (85,52,66), [140,146,173]
#> 159: (52,10,2), [166,150,173]
#> 160: (41,84,64), [142,161,0]
#> 161: (41,64,25), [155,162,160]
#> 162: (41,25,46), [155,165,161]
#> 163: (15,75,74), [171,170,151]
#> 164: (26,74,85), [175,146,170]
#> 165: (44,41,46), [162,157,179]
#> 166: (33,2,10), [159,181,180]
#> 167: (23,75,48), [151,177,184]
#> 168: (37,36,47), [172,169,130]
#> 169: (47,48,37), [138,168,153]
#> 170: (15,74,26), [164,147,163]
#> 171: (99,74,75), [163,184,183]
#> 172: (47,36,76), [114,132,168]
#> 173: (52,85,10), [174,159,158]
#> 174: (85,9,10), [181,173,175]
#> 175: (85,74,9), [183,174,164]
#> 176: (22,23,54), [177,188,185]
#> 177: (48,54,23), [176,167,153]
#> 178: (54,47,86), [132,0,153]
#> 179: (33,41,44), [165,180,0]
#> 180: (33,44,2), [156,166,179]
#> 181: (33,10,9), [174,182,166]
#> 182: (99,33,9), [181,183,186]
#> 183: (99,9,74), [175,171,182]
#> 184: (99,75,23), [167,185,171]
#> 185: (22,99,23), [184,176,187]
#> 186: (65,33,99), [182,187,0]
#> 187: (65,99,22), [185,188,186]
#> 188: (65,22,54), [176,0,187]
#> boundary nodes:  6 97 1 94 86 54 65 33 41 84 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.07384161 0.12281143 0.23036283 0.17325477 0.20694782 0.09108302 0.06208184 
#>                                                                              
#> 0.06908092 0.26507863 0.17252218 0.05542078 0.13323584 0.07996112 0.05976438 
#>                                                                              
#> 0.15848934 0.15936912 0.08267997 0.17223766 0.16344192 0.13050745 0.13061767 
#>                                                                              
#> 0.14171263 0.17245376 0.08436136 0.10783165 0.29742849 0.05347491 0.19784704 
#>                                                                              
#> 0.10860372 0.14949149 0.15705419 0.14634609 0.08345218 0.17124414 0.06524229 
#>                                                                              
#> 0.02856879 0.05211750 0.05667336 0.10284513 0.03889144 0.03444476 0.02288880 
#>                                                                              
#> 0.14081029 0.13158792 0.18590144 0.05053008 0.12729404 0.09887379 0.17510317 
#>                                                                              
#> 0.14363297 0.13046483 0.02191483 0.13903922 0.44924501 0.05653165 0.13777126 
#>                                                                              
#> 0.13976795 0.08065234 0.08911260 0.20699806 0.08467306 0.04608628 0.10519556 
#>                                                                              
#> 0.27328843 0.11530516 0.16775985 0.04024429 0.07631067 0.03691738 0.12801873 
#>                                                                              
#> 0.25051490 0.09206116 0.02420155 0.06083883 0.04191450 0.07199598 0.11768295 
#>                                                                              
#> 0.05983989 0.02843342 0.07336294 0.05313877 0.08694306 0.06971640 0.10783532 
#>                                                                              
#> 0.04448045 0.10332586 0.05871542 0.18156554 0.17909012 0.23845879 0.09051891 
#>                                                                              
#> 0.05005667 0.15583334 0.04909354 0.07584437 0.15070624 0.05246620 0.08702473 
#>                                                                              
#> 0.05651153 0.06793407 0.06189028 0.01659472 0.11128418 0.05080318 0.19205865 
#>                                                                              
#> 0.11675065 0.34688532 0.10109885 0.15484635 0.01117059 0.05406978 0.07186022 
#>                                                                              
#> 0.14337815 0.03883643 0.10421664 0.07208540 0.09956184 0.24803795 0.19748325 
#>                                                                              
#> 0.10278260 0.13141831 0.09923043 0.07521665 0.10539204 0.05195389 0.13919349 
#>                                                                              
#> 0.17704819 0.11104813 0.05750525 0.12793089 0.06104421 0.03446309 0.26683543 
#>                                                                              
#> 0.08031986 0.11693415 0.19172702 0.14792082 0.13388479 0.11233592 0.29168290 
#>                                                                              
#> 0.19358920 0.62165204 0.17488859 0.07650848 0.24139162 0.07638010 0.16863827 
#>                                                                              
#> 0.55228184 0.20734220 0.20547698 0.16600649 0.12930431 0.08759800 0.14796214 
#>                                                                              
#> 0.08511656 0.18266146 0.06964140 0.11685616 0.06033828 0.12835114 0.12934977 
#>                                                                              
#> 0.19363986 0.11352062 0.21713009 0.20142004 0.18542072 0.12792053 0.14706032 
#>                                                                              
#> 0.08500195 0.24394180 0.01332466 0.10989852 0.04199106 0.08693739 0.05668214 
#>                                                                              
#> 0.06385258 0.06265959 0.04762302 0.16684517 0.12568992 0.14902080 0.39196018 
#>                                                                              
#> 0.23150067 0.07873263 0.11274145 0.17990518 0.13626641 0.16502111 0.13817699 
#>                                                                              
#> 0.16338061 0.10815777 0.05035908 0.17366036 0.05837055 0.12809674 0.18380520 
#>                                                                              
#> 0.02198892 0.03890131 0.03191665 0.05936588 0.06028999 0.11640696 0.12514578 
#>                                                                              
#> 0.57419938 0.04420281 0.03996890 0.06008364 0.16326890 0.15912499 0.09338021 
#>                                                                              
#> 0.26485022 0.06753678 0.12985837 0.07556976 0.05118798 0.10946349 0.12603802 
#>                                                                              
#> 0.09135552 0.10772260 0.08121182 0.15296884 0.06705869 0.04046157 0.07436164 
#>                                                                              
#> 0.38966155 0.10422411 0.07399225 0.11864741 0.07544313 0.04621694 0.08121514 
#>                                                                              
#> 0.09708342 0.03731673 0.20813054 0.14104438 0.06746988 0.05744876 0.18214930 
#>                                                                              
#> 0.08840231 0.17046094 0.20193124 0.11220049 0.24760412 0.12659472 0.14042512 
#>                                                                              
#> 0.02134010 0.13847610 0.12739207 0.13012243 0.08018104 0.17720603 0.13034748 
#>                                                                              
#> 0.31777167 0.17295050 0.06106543 0.67291442 0.06791483 0.09886898 0.31109281 
#>                                                                              
#> 0.10824007 0.19872990 0.20147924 0.03800685 0.24346199 0.05798557 0.04232347 
#>                                                                              
#> 0.03055329 0.09457958 0.03377781 0.12777585 0.11023518 0.13990310 0.17519393 
#>                                                                              
#> 0.14338192 0.18546350 0.08955136 0.16806723 0.09980026 0.04254761 0.58108980 
#>                                                                              
#> 0.09007267 0.12079174 0.20260506 0.17736115 0.19760460 0.23147819 0.75356875 
#> 
#> $alpha
#> [1] 0.3166676
#> 
#> attr(,"class")
#> [1] "scree"