Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

Usage,

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2         mx1
#>    86   36 0.923201170 0.932654111 0.805190030 0.856631952  0.83512091
#>    36   78 0.805190030 0.856631952 0.834387504 0.821410167  0.85757373
#>    78   24 0.834387504 0.821410167 0.855456930 0.766541144  0.90656037
#>    24   77 0.855456930 0.766541144 0.932233908 0.583334996  1.06903048
#>    77   17 0.932233908 0.583334996 1.000000000 0.508479259  1.20125725
#>    17   86 1.000000000 0.508479259 0.923201170 0.932654111  1.28763902
#>    91   72 0.174981115 0.128153323 0.235056049 0.109748577  0.23258989
#>    72   88 0.235056049 0.109748577 0.293968755 0.108967836  0.26516965
#>    88   38 0.293968755 0.108967836 0.373265849 0.171540335  0.35294533
#>    38   42 0.373265849 0.171540335 0.349119741 0.276260084  0.34292676
#>    42   13 0.349119741 0.276260084 0.322223036 0.303261384  0.28060699
#>    13   10 0.322223036 0.303261384 0.254234404 0.321960222  0.26685531
#>    10   91 0.254234404 0.321960222 0.174981115 0.128153323  0.23003003
#>    19   25 0.173870063 0.763321330 0.114964496 0.638718470  0.16342096
#>    25   50 0.114964496 0.638718470 0.173518891 0.613417420  0.16723349
#>    50   73 0.173518891 0.613417420 0.290617834 0.587243375  0.23484264
#>    73   69 0.290617834 0.587243375 0.297672549 0.611222038  0.23896246
#>    69    8 0.297672549 0.611222038 0.298827561 0.747893915  0.30137586
#>     8   19 0.298827561 0.747893915 0.173870063 0.763321330  0.23431575
#>    95   83 0.485637254 1.000000000 0.597462516 0.993480659  0.56683251
#>    83   61 0.597462516 0.993480659 0.686387284 0.971380579  0.62748429
#>    61   96 0.686387284 0.971380579 0.933711302 0.956724087  0.82777324
#>    65   70 0.884836496 0.596521865 0.884858146 0.599714490  0.87838790
#>    70   24 0.884858146 0.599714490 0.855456930 0.766541144  0.90609620
#>    24   74 0.855456930 0.766541144 0.687070928 0.623132046  0.78724296
#>    74   23 0.687070928 0.623132046 0.735031567 0.524636989  0.77985845
#>    23    7 0.735031567 0.524636989 0.800626542 0.506171667  0.78911121
#>     7   65 0.800626542 0.506171667 0.884836496 0.596521865  0.84774118
#>    99   97 0.005444609 0.525236220 0.000000000 0.416087142  0.04618999
#>    97   91 0.000000000 0.416087142 0.174981115 0.128153323  0.04520216
#>    91   98 0.174981115 0.128153323 0.175284317 0.366736710  0.04929998
#>    98   11 0.175284317 0.366736710 0.157053021 0.411095057  0.12146812
#>    11   18 0.157053021 0.411095057 0.125128759 0.461328722  0.12538621
#>    18   99 0.125128759 0.461328722 0.005444609 0.525236220  0.05631860
#>    23   34 0.735031567 0.524636989 0.737342499 0.382756278  0.75052376
#>    34   16 0.737342499 0.382756278 0.847814258 0.341312416  0.81470664
#>    16   71 0.847814258 0.341312416 0.932608006 0.453384951  0.84678889
#>    71   65 0.932608006 0.453384951 0.884836496 0.596521865  0.88098249
#>    76   57 0.384422070 0.696787419 0.446003844 0.726066401  0.37970277
#>    57   82 0.446003844 0.726066401 0.323271266 0.873486489  0.38131921
#>    82   37 0.323271266 0.873486489 0.158308574 0.847189367  0.24670814
#>    37   19 0.158308574 0.847189367 0.173870063 0.763321330  0.24426933
#>    69   76 0.297672549 0.611222038 0.384422070 0.696787419  0.31599075
#>    45   72 0.333155970 0.007990054 0.235056049 0.109748577  0.24464473
#>    72   27 0.235056049 0.109748577 0.148787045 0.055937952  0.21922791
#>    27   28 0.148787045 0.055937952 0.131713276 0.000000000  0.14304044
#>    28   45 0.131713276 0.000000000 0.333155970 0.007990054  0.44390432
#>    13   81 0.322223036 0.303261384 0.281446633 0.361609877  0.29174108
#>    81   53 0.281446633 0.361609877 0.204579468 0.401426928  0.23434077
#>    53   98 0.204579468 0.401426928 0.175284317 0.366736710  0.22195620
#>    53   43 0.204579468 0.401426928 0.191228118 0.495591275  0.18915837
#>    43   18 0.191228118 0.495591275 0.125128759 0.461328722  0.17041711
#>    32  100 0.973647643 0.034272737 0.628609022 0.114874267  0.72592844
#>   100   39 0.628609022 0.114874267 0.594586989 0.139307998  0.57247497
#>    39   26 0.594586989 0.139307998 0.484654057 0.177237135  0.52693664
#>    26   63 0.484654057 0.177237135 0.438338360 0.109797000  0.49990759
#>    63   93 0.438338360 0.109797000 0.448189122 0.080303589  0.49629811
#>    93   55 0.448189122 0.080303589 0.444120176 0.053777016  0.47404349
#>    55   15 0.444120176 0.053777016 0.440217475 0.033234417  0.47468362
#>    15   32 0.440217475 0.033234417 0.973647643 0.034272737  0.70824798
#>    42   89 0.349119741 0.276260084 0.461612678 0.349593602  0.38708297
#>    89   90 0.461612678 0.349593602 0.396490928 0.428609974  0.38355021
#>    90   81 0.396490928 0.428609974 0.281446633 0.361609877  0.35416942
#>    94   85 0.554216826 0.204615443 0.581180342 0.243702207  0.54070904
#>    85   89 0.581180342 0.243702207 0.461612678 0.349593602  0.52447888
#>    38   26 0.373265849 0.171540335 0.484654057 0.177237135  0.42600539
#>    26   94 0.484654057 0.177237135 0.554216826 0.204615443  0.52062255
#>    45   32 0.333155970 0.007990054 0.973647643 0.034272737  0.70962402
#>    55   45 0.444120176 0.053777016 0.333155970 0.007990054  0.37843350
#>    32   64 0.973647643 0.034272737 0.957669620 0.314809390  0.86431212
#>    64   48 0.957669620 0.314809390 0.977394259 0.359967915  0.91095425
#>    48   60 0.977394259 0.359967915 0.972475208 0.394673374  0.91008638
#>    60   71 0.972475208 0.394673374 0.932608006 0.453384951  0.90110813
#>    34   51 0.737342499 0.382756278 0.731173977 0.318549644  0.78642829
#>    51   32 0.731173977 0.318549644 0.973647643 0.034272737  0.82415707
#>    96   86 0.933711302 0.956724087 0.923201170 0.932654111  1.30100274
#>    77   71 0.932233908 0.583334996 0.932608006 0.453384951  0.93569826
#>    48   32 0.977394259 0.359967915 0.973647643 0.034272737  2.67479779
#>    43   67 0.191228118 0.495591275 0.079254404 0.594860651  0.12787646
#>    67   99 0.079254404 0.594860651 0.005444609 0.525236220  0.07976047
#>    37   20 0.158308574 0.847189367 0.151682555 0.822910980  0.22447347
#>    20   58 0.151682555 0.822910980 0.136303890 0.790898493  0.16836453
#>    58   25 0.136303890 0.790898493 0.114964496 0.638718470  0.11087142
#>    37   21 0.158308574 0.847189367 0.119922538 0.838089070  0.13994064
#>    21   58 0.119922538 0.838089070 0.136303890 0.790898493  0.12816112
#>    37   35 0.158308574 0.847189367 0.103926275 0.885609174  0.13288840
#>    35   99 0.103926275 0.885609174 0.005444609 0.525236220 -0.17020229
#>    67   58 0.079254404 0.594860651 0.136303890 0.790898493 -0.07173445
#>    78   56 0.834387504 0.821410167 0.617807994 0.783406831  0.73485570
#>    56   87 0.617807994 0.783406831 0.627940521 0.654887113  0.65374104
#>    87   74 0.627940521 0.654887113 0.687070928 0.623132046  0.64484067
#>    74   40 0.687070928 0.623132046 0.711616700 0.528385086  0.73258589
#>    40   34 0.711616700 0.528385086 0.737342499 0.382756278  0.71160439
#>    70   77 0.884858146 0.599714490 0.932233908 0.583334996  0.94476228
#>    67   50 0.079254404 0.594860651 0.173518891 0.613417420  0.12900944
#>    38   63 0.373265849 0.171540335 0.438338360 0.109797000  0.42941192
#>    39   94 0.594586989 0.139307998 0.554216826 0.204615443  0.53619491
#>    91   79 0.174981115 0.128153323 0.134493312 0.067996260  0.16581257
#>    79   28 0.134493312 0.067996260 0.131713276 0.000000000  0.11854737
#>    79   59 0.134493312 0.067996260 0.070676366 0.052331581  0.10876329
#>    59   75 0.070676366 0.052331581 0.032836275 0.055113963  0.04490232
#>    62   46 0.517900567 0.579270020 0.500755000 0.621909591  0.49996871
#>    46   54 0.500755000 0.621909591 0.493711121 0.641342690  0.48074425
#>    54   76 0.493711121 0.641342690 0.384422070 0.696787419  0.42958145
#>    76   30 0.384422070 0.696787419 0.328188098 0.573178890  0.39635454
#>    30   49 0.328188098 0.573178890 0.363038885 0.531183428  0.40052764
#>    49   41 0.363038885 0.531183428 0.480224336 0.450864277  0.44666008
#>    41   62 0.480224336 0.450864277 0.517900567 0.579270020  0.45580446
#>    73   52 0.290617834 0.587243375 0.277798421 0.522260122  0.29879498
#>    52   49 0.277798421 0.522260122 0.363038885 0.531183428  0.31998032
#>    41   90 0.480224336 0.450864277 0.396490928 0.428609974  0.44109400
#>    85   84 0.581180342 0.243702207 0.720084143 0.365871210  0.59273590
#>    84   40 0.720084143 0.365871210 0.711616700 0.528385086  0.61468774
#>    40   62 0.711616700 0.528385086 0.517900567 0.579270020  0.59703205
#>    51  100 0.731173977 0.318549644 0.628609022 0.114874267  0.81807669
#>    95   44 0.485637254 1.000000000 0.363397828 0.952155291  0.30297581
#>    44   82 0.363397828 0.952155291 0.323271266 0.873486489  0.42355307
#>    57   56 0.446003844 0.726066401 0.617807994 0.783406831  0.49924148
#>    56   95 0.617807994 0.783406831 0.485637254 1.000000000  0.51788858
#>    84   51 0.720084143 0.365871210 0.731173977 0.318549644  0.75368870
#>    37   66 0.158308574 0.847189367 0.170577054 0.951261934  0.15508761
#>    86   92 0.923201170 0.932654111 0.754531530 0.882728004  0.82447952
#>    92   56 0.754531530 0.882728004 0.617807994 0.783406831  0.72995267
#>    82   66 0.323271266 0.873486489 0.170577054 0.951261934  0.23593421
#>    88   63 0.293968755 0.108967836 0.438338360 0.109797000  0.36621350
#>   100   85 0.628609022 0.114874267 0.581180342 0.243702207  0.66496793
#>    74   62 0.687070928 0.623132046 0.517900567 0.579270020  0.61476366
#>    49   90 0.363038885 0.531183428 0.396490928 0.428609974  0.42386200
#>    53   80 0.204579468 0.401426928 0.310795500 0.468541140  0.24492849
#>    80   52 0.310795500 0.468541140 0.277798421 0.522260122  0.24752947
#>    52   50 0.277798421 0.522260122 0.173518891 0.613417420  0.21878635
#>    95   66 0.485637254 1.000000000 0.170577054 0.951261934  0.26498613
#>    55   63 0.444120176 0.053777016 0.438338360 0.109797000  0.37064007
#>    62   68 0.517900567 0.579270020 0.536196151 0.651066264  0.53839325
#>    68   54 0.536196151 0.651066264 0.493711121 0.641342690  0.51667714
#>    68   56 0.536196151 0.651066264 0.617807994 0.783406831  0.56942748
#>    57   54 0.446003844 0.726066401 0.493711121 0.641342690  0.44613141
#>    80   90 0.310795500 0.468541140 0.396490928 0.428609974  0.36617220
#>    81   80 0.281446633 0.361609877 0.310795500 0.468541140  0.26484251
#>    92   61 0.754531530 0.882728004 0.686387284 0.971380579  0.65464444
#>    68   87 0.536196151 0.651066264 0.627940521 0.654887113  0.57945513
#>    91   75 0.174981115 0.128153323 0.032836275 0.055113963  0.06072367
#>    92   96 0.754531530 0.882728004 0.933711302 0.956724087  0.81202785
#>     1   36 0.908067292 0.889899719 0.805190030 0.856631952  0.83512091
#>     1   78 0.908067292 0.889899719 0.834387504 0.821410167  0.85757373
#>     1   24 0.908067292 0.889899719 0.855456930 0.766541144  0.90656037
#>     1   77 0.908067292 0.889899719 0.932233908 0.583334996  1.06903048
#>     1   17 0.908067292 0.889899719 1.000000000 0.508479259  1.20125725
#>     1   86 0.908067292 0.889899719 0.923201170 0.932654111  1.28763902
#>     2   72 0.322343710 0.166631506 0.235056049 0.109748577  0.23258989
#>     2   88 0.322343710 0.166631506 0.293968755 0.108967836  0.26516965
#>     2   38 0.322343710 0.166631506 0.373265849 0.171540335  0.35294533
#>     2   42 0.322343710 0.166631506 0.349119741 0.276260084  0.34292676
#>     2   13 0.322343710 0.166631506 0.322223036 0.303261384  0.28060699
#>     2   10 0.322343710 0.166631506 0.254234404 0.321960222  0.26685531
#>     2   91 0.322343710 0.166631506 0.174981115 0.128153323  0.23003003
#>     3   25 0.233186218 0.674047236 0.114964496 0.638718470  0.16342096
#>     3   50 0.233186218 0.674047236 0.173518891 0.613417420  0.16723349
#>     3   73 0.233186218 0.674047236 0.290617834 0.587243375  0.23484264
#>     3   69 0.233186218 0.674047236 0.297672549 0.611222038  0.23896246
#>     3    8 0.233186218 0.674047236 0.298827561 0.747893915  0.30137586
#>     3   19 0.233186218 0.674047236 0.173870063 0.763321330  0.23431575
#>     4   83 0.640314733 0.998616298 0.597462516 0.993480659  0.56683251
#>     4   61 0.640314733 0.998616298 0.686387284 0.971380579  0.62748429
#>     4   96 0.640314733 0.998616298 0.933711302 0.956724087  0.82777324
#>     5   70 0.873871870 0.603048689 0.884858146 0.599714490  0.87838790
#>     5   24 0.873871870 0.603048689 0.855456930 0.766541144  0.90609620
#>     5   74 0.873871870 0.603048689 0.687070928 0.623132046  0.78724296
#>     5   23 0.873871870 0.603048689 0.735031567 0.524636989  0.77985845
#>     5    7 0.873871870 0.603048689 0.800626542 0.506171667  0.78911121
#>     5   65 0.873871870 0.603048689 0.884836496 0.596521865  0.84774118
#>     6   97 0.091911938 0.415678295 0.000000000 0.416087142  0.04618999
#>     6   91 0.091911938 0.415678295 0.174981115 0.128153323  0.04520216
#>     6   98 0.091911938 0.415678295 0.175284317 0.366736710  0.04929998
#>     6   11 0.091911938 0.415678295 0.157053021 0.411095057  0.12146812
#>     6   18 0.091911938 0.415678295 0.125128759 0.461328722  0.12538621
#>     6   99 0.091911938 0.415678295 0.005444609 0.525236220  0.05631860
#>     7   34 0.800626542 0.506171667 0.737342499 0.382756278  0.75052376
#>     7   16 0.800626542 0.506171667 0.847814258 0.341312416  0.81470664
#>     7   71 0.800626542 0.506171667 0.932608006 0.453384951  0.84678889
#>     8   57 0.298827561 0.747893915 0.446003844 0.726066401  0.37970277
#>     8   82 0.298827561 0.747893915 0.323271266 0.873486489  0.38131921
#>     8   37 0.298827561 0.747893915 0.158308574 0.847189367  0.24670814
#>     8   76 0.298827561 0.747893915 0.384422070 0.696787419  0.31599075
#>     9   72 0.157334033 0.001439569 0.235056049 0.109748577  0.24464473
#>     9   27 0.157334033 0.001439569 0.148787045 0.055937952  0.21922791
#>     9   28 0.157334033 0.001439569 0.131713276 0.000000000  0.14304044
#>     9   45 0.157334033 0.001439569 0.333155970 0.007990054  0.44390432
#>    10   81 0.254234404 0.321960222 0.281446633 0.361609877  0.29174108
#>    10   53 0.254234404 0.321960222 0.204579468 0.401426928  0.23434077
#>    10   98 0.254234404 0.321960222 0.175284317 0.366736710  0.22195620
#>    11   53 0.157053021 0.411095057 0.204579468 0.401426928  0.17830226
#>    11   43 0.157053021 0.411095057 0.191228118 0.495591275  0.18915837
#>    12  100 0.504380923 0.055293699 0.628609022 0.114874267  0.72592844
#>    12   39 0.504380923 0.055293699 0.594586989 0.139307998  0.57247497
#>    12   26 0.504380923 0.055293699 0.484654057 0.177237135  0.52693664
#>    12   63 0.504380923 0.055293699 0.438338360 0.109797000  0.49990759
#>    12   93 0.504380923 0.055293699 0.448189122 0.080303589  0.49629811
#>    12   55 0.504380923 0.055293699 0.444120176 0.053777016  0.47404349
#>    12   15 0.504380923 0.055293699 0.440217475 0.033234417  0.47468362
#>    12   32 0.504380923 0.055293699 0.973647643 0.034272737  0.70824798
#>    13   89 0.322223036 0.303261384 0.461612678 0.349593602  0.38708297
#>    13   90 0.322223036 0.303261384 0.396490928 0.428609974  0.38355021
#>    14   85 0.505754792 0.222357066 0.581180342 0.243702207  0.54070904
#>    14   89 0.505754792 0.222357066 0.461612678 0.349593602  0.52447888
#>    14   42 0.505754792 0.222357066 0.349119741 0.276260084  0.43414051
#>    14   38 0.505754792 0.222357066 0.373265849 0.171540335  0.42364933
#>    14   26 0.505754792 0.222357066 0.484654057 0.177237135  0.42600539
#>    14   94 0.505754792 0.222357066 0.554216826 0.204615443  0.52062255
#>    15   45 0.440217475 0.033234417 0.333155970 0.007990054  0.37843350
#>    16   64 0.847814258 0.341312416 0.957669620 0.314809390  0.86431212
#>    16   48 0.847814258 0.341312416 0.977394259 0.359967915  0.91095425
#>    16   60 0.847814258 0.341312416 0.972475208 0.394673374  0.91008638
#>    16   51 0.847814258 0.341312416 0.731173977 0.318549644  0.78642829
#>    16   32 0.847814258 0.341312416 0.973647643 0.034272737  0.82415707
#>    17   71 1.000000000 0.508479259 0.932608006 0.453384951  0.93569826
#>    17   60 1.000000000 0.508479259 0.972475208 0.394673374  0.99130486
#>    17   48 1.000000000 0.508479259 0.977394259 0.359967915  1.17563981
#>    17   32 1.000000000 0.508479259 0.973647643 0.034272737  2.67479779
#>    18   67 0.125128759 0.461328722 0.079254404 0.594860651  0.12787646
#>    19   20 0.173870063 0.763321330 0.151682555 0.822910980  0.22447347
#>    19   58 0.173870063 0.763321330 0.136303890 0.790898493  0.16836453
#>    20   21 0.151682555 0.822910980 0.119922538 0.838089070  0.13994064
#>    21   35 0.119922538 0.838089070 0.103926275 0.885609174  0.13288840
#>    21   99 0.119922538 0.838089070 0.005444609 0.525236220 -0.17020229
#>    21   67 0.119922538 0.838089070 0.079254404 0.594860651 -0.14355555
#>    22   24 0.668075771 0.651522026 0.855456930 0.766541144  0.76845535
#>    22   78 0.668075771 0.651522026 0.834387504 0.821410167  0.73543285
#>    22   56 0.668075771 0.651522026 0.617807994 0.783406831  0.73485570
#>    22   87 0.668075771 0.651522026 0.627940521 0.654887113  0.65374104
#>    22   74 0.668075771 0.651522026 0.687070928 0.623132046  0.64484067
#>    23   40 0.735031567 0.524636989 0.711616700 0.528385086  0.73258589
#>    25   67 0.114964496 0.638718470 0.079254404 0.594860651 -0.05425389
#>    27   91 0.148787045 0.055937952 0.174981115 0.128153323  0.19328624
#>    27   79 0.148787045 0.055937952 0.134493312 0.067996260  0.16581257
#>    28   59 0.131713276 0.000000000 0.070676366 0.052331581  0.10876329
#>    28   75 0.131713276 0.000000000 0.032836275 0.055113963  0.04490232
#>    29   46 0.476508738 0.605737730 0.500755000 0.621909591  0.49996871
#>    29   54 0.476508738 0.605737730 0.493711121 0.641342690  0.48074425
#>    29   76 0.476508738 0.605737730 0.384422070 0.696787419  0.42958145
#>    29   30 0.476508738 0.605737730 0.328188098 0.573178890  0.39635454
#>    29   49 0.476508738 0.605737730 0.363038885 0.531183428  0.40052764
#>    29   41 0.476508738 0.605737730 0.480224336 0.450864277  0.44666008
#>    29   62 0.476508738 0.605737730 0.517900567 0.579270020  0.45580446
#>    30   52 0.328188098 0.573178890 0.277798421 0.522260122  0.29879498
#>    30   69 0.328188098 0.573178890 0.297672549 0.611222038  0.36266155
#>    30   73 0.328188098 0.573178890 0.290617834 0.587243375  0.31430354
#>    31   41 0.484802776 0.438367039 0.480224336 0.450864277  0.59665486
#>    31   90 0.484802776 0.438367039 0.396490928 0.428609974  0.44109400
#>    31   89 0.484802776 0.438367039 0.461612678 0.349593602  0.44417392
#>    31   85 0.484802776 0.438367039 0.581180342 0.243702207  0.58234680
#>    31   84 0.484802776 0.438367039 0.720084143 0.365871210  0.59273590
#>    31   40 0.484802776 0.438367039 0.711616700 0.528385086  0.61468774
#>    31   62 0.484802776 0.438367039 0.517900567 0.579270020  0.59703205
#>    33   44 0.431300786 0.971898452 0.363397828 0.952155291  0.30297581
#>    33   82 0.431300786 0.971898452 0.323271266 0.873486489  0.42355307
#>    33   57 0.431300786 0.971898452 0.446003844 0.726066401  0.44413515
#>    33   56 0.431300786 0.971898452 0.617807994 0.783406831  0.49924148
#>    33   95 0.431300786 0.971898452 0.485637254 1.000000000  0.51788858
#>    34   84 0.737342499 0.382756278 0.720084143 0.365871210  0.66030531
#>    35   66 0.103926275 0.885609174 0.170577054 0.951261934  0.15508761
#>    36   92 0.805190030 0.856631952 0.754531530 0.882728004  0.82447952
#>    36   56 0.805190030 0.856631952 0.617807994 0.783406831  0.72995267
#>    39   85 0.594586989 0.139307998 0.581180342 0.243702207  0.66496793
#>    43   80 0.191228118 0.495591275 0.310795500 0.468541140  0.24492849
#>    43   52 0.191228118 0.495591275 0.277798421 0.522260122  0.24752947
#>    43   50 0.191228118 0.495591275 0.173518891 0.613417420  0.21878635
#>    44   66 0.363397828 0.952155291 0.170577054 0.951261934  0.26498613
#>    45   63 0.333155970 0.007990054 0.438338360 0.109797000  0.37064007
#>    45   88 0.333155970 0.007990054 0.293968755 0.108967836  0.36632831
#>    46   68 0.500755000 0.621909591 0.536196151 0.651066264  0.53839325
#>    47   56 0.499140661 0.687518906 0.617807994 0.783406831  0.56942748
#>    47   57 0.499140661 0.687518906 0.446003844 0.726066401  0.52416912
#>    47   54 0.499140661 0.687518906 0.493711121 0.641342690  0.44613141
#>    47   68 0.499140661 0.687518906 0.536196151 0.651066264  0.51117920
#>    49   80 0.363038885 0.531183428 0.310795500 0.468541140  0.32192117
#>    51   85 0.731173977 0.318549644 0.581180342 0.243702207  0.63615859
#>    56   61 0.617807994 0.783406831 0.686387284 0.971380579  0.65464444
#>    56   83 0.617807994 0.783406831 0.597462516 0.993480659  0.61883636
#>    59   91 0.070676366 0.052331581 0.174981115 0.128153323  0.08111665
#>    62   87 0.517900567 0.579270020 0.627940521 0.654887113  0.61378903
#>    65   77 0.884836496 0.596521865 0.932233908 0.583334996  0.88858838
#>    75   97 0.032836275 0.055113963 0.000000000 0.416087142  0.02932958
#>     4   95 0.640314733 0.998616298 0.485637254 1.000000000  0.56683251
#>    17   96 1.000000000 0.508479259 0.933711302 0.956724087  1.30100274
#>          my1         mx2          my2 bp1 bp2
#>   0.93977641  0.82447952   0.95629532   0   0
#>   0.87034331  0.73238667   0.76656816   0   0
#>   0.81764439  0.73543285   0.75193230   0   0
#>   0.74835362  0.94476228   0.69627602   0   0
#>   0.75877710  0.93569826   0.51836941   0   0
#>   0.77959746  1.30100274   0.78201702   0   0
#>   0.20894647  0.19328624   0.08065540   0   0
#>   0.15895237  0.26358107   0.03908233   0   0
#>   0.11575999  0.36621350   0.09894546   0   0
#>   0.21968846  0.42364933   0.23830134   0   0
#>   0.23490964  0.38708297   0.34097316   0   0
#>   0.23489749  0.29174108   0.32538169   0   0
#>   0.21875016  0.15981168   0.24746449   0   0
#>   0.69203597  0.11087142   0.71687859   0   0
#>   0.67927801  0.12900944   0.59081583   0   0
#>   0.61274214  0.22451282   0.56652799   0   0
#>   0.61546791  0.31430354   0.59330196   0   0
#>   0.67953156  0.31599075   0.67940805   0   0
#>   0.73914041  0.24426933   0.81976144   0   0
#>   1.43040926  0.53482516   0.88139215   0   0
#>   0.92432550  0.61883636   0.88952856   0   0
#>   1.26313875  0.81202785   0.99743987   0   0
#>   0.59816198  0.91076485   0.59794243   0   0
#>   0.68946158  0.94476228   0.69627602   0   0
#>   0.67607457  0.76845535   0.69813433   0   0
#>   0.60738911  0.73258589   0.58437048   0   0
#>   0.59100563  0.75052376   0.45393015   0   0
#>   0.54667756  0.88098249   0.51569535   0   0
#>   0.46849341 -9.09285126   0.92436994   0   1
#>   0.24642103  0.02932958   0.23677505   0   0
#>   0.24760493  0.15981168   0.24746449   0   0
#>   0.37054394  0.17830226   0.39390279   0   0
#>   0.42623133  0.17041711   0.45484915   0   0
#>   0.47648730  0.07976047   0.52038857   0   0
#>   0.45393015  0.71160439   0.45329623   0   0
#>   0.42101892  0.78642829   0.34564083   0   0
#>   0.43020183  0.90110813   0.38910405   0   0
#>   0.51569535  0.88858838   0.51823379   0   0
#>   0.78611463  0.43715002   0.66528722   0   0
#>   0.79701380  0.44413515   0.84931035   0   0
#>   0.82321275  0.23593421   0.89079797   0   0
#>   0.81976144  0.22447347   0.81608837   0   0
#>   0.67940805  0.36266155   0.63209137   0   0
#>   0.02082681  0.26358107   0.03908233   0   0
#>   0.03906580  0.19328624   0.08065540   0   0
#>   0.02711731  0.11854737   0.03459325   0   0
#>  -5.32751153  0.64049547 -10.28390486   0   1
#>   0.32538169  0.35416942   0.36900926   0   0
#>   0.36477649  0.26484251   0.42366037   0   0
#>   0.35703797  0.17830226   0.39390279   0   0
#>   0.44726911  0.24492849   0.45517663   0   0
#>   0.45484915  0.12787646   0.53691862   0   0
#>  -0.24734180  0.81807669   0.14712594   0   0
#>   0.07261542  0.66496793   0.20140452   0   0
#>   0.12150988  0.53619491   0.14834385   0   0
#>   0.11713738  0.42941192   0.16555152   0   0
#>   0.11276370  0.38670421   0.07615948   0   0
#>   0.06276240  0.38670421   0.07615948   0   0
#>   0.03732853  0.37843350   0.05561421   0   0
#>  -0.64203604  0.70962402  -1.34896692   0   0
#>   0.34097316  0.43414051   0.26878733   0   0
#>   0.35160141  0.44417392   0.40156476   0   0
#>   0.36900926  0.33328091   0.40487644   0   0
#>   0.24277723  0.61077282   0.19444460   0   0
#>   0.30012838  0.58234680   0.36547015   0   0
#>   0.23215864  0.42941192   0.16555152   0   0
#>   0.18791009  0.53619491   0.14834385   0   0
#>  -1.34896692  0.77358455  -2.90764284   0   1
#>   0.05561421  0.37064007   0.07450151   0   0
#>   0.16876885  1.29713921   0.19342061   0   0
#>   0.36210103  1.29713921   0.19342061   0   0
#>   0.36812923  1.17563981   0.40576800   0   0
#>   0.38910405  0.99130486   0.45035076   0   0
#>   0.34564083  0.75368870   0.34878622   0   0
#>   0.15231220  0.81807669   0.14712594   0   0
#>   0.78201702  0.81299137   0.99510672   0   0
#>   0.51836941  0.88858838   0.51823379   0   0
#>   0.17757279  1.29713921   0.19342061   0   0
#>   0.53691862  0.13748579   0.54775774   0   0
#>   0.52038857 -0.14355555   0.75712888   0   0
#>   0.81608837  0.13994064   0.83915894   0   0
#>   0.79519686  0.12816112   0.81451041   0   0
#>   0.71687859 -0.05425389   0.74003323   0   0
#>   0.83915894  0.13288840   0.86890604   0   0
#>   0.81451041 -0.07173445   0.74512029   0   0
#>   0.86890604  0.15508761   0.90032848   0   0
#>   0.76687936 -2.75227519   1.47250069   0   1
#>   0.74512029 -0.05425389   0.74003323   0   0
#>   0.75249730  0.73238667   0.76656816   0   0
#>   0.72158052  0.57945513   0.71572380   0   0
#>   0.61542627  0.61378903   0.55760569   0   0
#>   0.58437048  0.61476366   0.55384667   0   0
#>   0.45329623  0.66030531   0.44423409   0   0
#>   0.69627602  0.91076485   0.59794243   0   0
#>   0.59081583  0.13748579   0.54775774   0   0
#>   0.16555152  0.36621350   0.09894546   0   0
#>   0.14834385  0.61077282   0.19444460   0   0
#>   0.09062069  0.08111665   0.14762400   0   0
#>   0.03459325  0.10876329   0.03499327   0   0
#>   0.03499327  0.08111665   0.14762400   0   0
#>  -0.03949094  0.06072367   0.17567776   0   0
#>   0.59682648  0.53839325   0.61227716   0   0
#>   0.62564947  0.51667714   0.63867400   0   0
#>   0.65036856  0.43715002   0.66528722   0   0
#>   0.61676322  0.36266155   0.63209137   0   0
#>   0.59775278  0.31998032   0.53090895   0   0
#>   0.52754033  0.42386200   0.49427798   0   0
#>   0.52775971  0.59665486   0.48643203   0   0
#>   0.55187416  0.22451282   0.56652799   0   0
#>   0.53090895  0.32192117   0.51236890   0   0
#>   0.42944135  0.42386200   0.49427798   0   0
#>   0.37061374  0.63615859   0.32124297   0   0
#>   0.44185728  0.66030531   0.44423409   0   0
#>   0.48634343  0.61476366   0.55384667   0   0
#>   0.14712594  0.69114876   0.21104313   0   0
#>   1.28660711  0.26498613   1.38366771   0   0
#>   0.87190386  0.26698719   0.95176335   0   0
#>   0.85260623  0.52416912   0.77791772   0   0
#>   0.87105701  0.53482516   0.88139215   0   0
#>   0.34878622  0.63615859   0.32124297   0   0
#>   0.90032848  0.23593421   0.89079797   0   0
#>   0.95629532  0.81299137   0.99510672   0   0
#>   0.77279674  0.65464444   0.87646454   0   0
#>   0.89079797  0.26698719   0.95176335   0   0
#>   0.09894546  0.36632831   0.07895623   0   0
#>   0.20140452  0.69114876   0.21104313   0   0
#>   0.55384667  0.61378903   0.55760569   0   0
#>   0.49427798  0.36617220   0.47546374   0   0
#>   0.45517663  0.26484251   0.42366037   0   0
#>   0.46667353  0.32192117   0.51236890   0   0
#>   0.55997718  0.22451282   0.56652799   0   0
#>   1.38366771 -0.21453834   4.48348452   0   1
#>   0.07450151  0.38670421   0.07615948   0   0
#>   0.61227716  0.58425000   0.60059165   0   0
#>   0.63867400  0.51117920   0.66269606   0   0
#>   0.72190765  0.57945513   0.71572380   0   0
#>   0.67034458  0.43715002   0.66528722   0   0
#>   0.47546374  0.33328091   0.40487644   0   0
#>   0.42366037  0.33328091   0.40487644   0   0
#>   0.87646454  0.81202785   0.99743987   0   0
#>   0.71572380  0.58425000   0.60059165   0   0
#>   0.17567776  0.02932958   0.23677505   0   0
#>   0.99743987  0.81299137   0.99510672   0   0
#>   0.93977641  0.85757373   0.87034331   0   0
#>   0.87034331  0.90656037   0.81764439   0   0
#>   0.81764439  1.06903048   0.74835362   0   0
#>   0.74835362  1.20125725   0.75877710   0   0
#>   0.75877710  1.28763902   0.77959746   0   0
#>   0.77959746  0.83512091   0.93977641   0   0
#>   0.20894647  0.26516965   0.15895237   0   0
#>   0.15895237  0.35294533   0.11575999   0   0
#>   0.11575999  0.34292676   0.21968846   0   0
#>   0.21968846  0.28060699   0.23490964   0   0
#>   0.23490964  0.26685531   0.23489749   0   0
#>   0.23489749  0.23003003   0.21875016   0   0
#>   0.21875016  0.23258989   0.20894647   0   0
#>   0.69203597  0.16723349   0.67927801   0   0
#>   0.67927801  0.23484264   0.61274214   0   0
#>   0.61274214  0.23896246   0.61546791   0   0
#>   0.61546791  0.30137586   0.67953156   0   0
#>   0.67953156  0.23431575   0.73914041   0   0
#>   0.73914041  0.16342096   0.69203597   0   0
#>   1.43040926  0.62748429   0.92432550   0   0
#>   0.92432550  0.82777324   1.26313875   0   0
#>   1.26313875  1.30470719   4.60339657   0   1
#>   0.59816198  0.90609620   0.68946158   0   0
#>   0.68946158  0.78724296   0.67607457   0   0
#>   0.67607457  0.77985845   0.60738911   0   0
#>   0.60738911  0.78911121   0.59100563   0   0
#>   0.59100563  0.84774118   0.54667756   0   0
#>   0.54667756  0.87838790   0.59816198   0   0
#>   0.46849341  0.04520216   0.24642103   0   0
#>   0.24642103  0.04929998   0.24760493   0   0
#>   0.24760493  0.12146812   0.37054394   0   0
#>   0.37054394  0.12538621   0.42623133   0   0
#>   0.42623133  0.05631860   0.47648730   0   0
#>   0.47648730  0.04618999   0.46849341   0   0
#>   0.45393015  0.81470664   0.42101892   0   0
#>   0.42101892  0.84678889   0.43020183   0   0
#>   0.43020183  0.88098249   0.51569535   0   0
#>   0.78611463  0.38131921   0.79701380   0   0
#>   0.79701380  0.24670814   0.82321275   0   0
#>   0.82321275  0.24426933   0.81976144   0   0
#>   0.67940805  0.37970277   0.78611463   0   0
#>   0.02082681  0.21922791   0.03906580   0   0
#>   0.03906580  0.14304044   0.02711731   0   0
#>   0.02711731  0.44390432  -5.32751153   0   0
#>  -5.32751153  0.24464473   0.02082681   0   0
#>   0.32538169  0.23434077   0.36477649   0   0
#>   0.36477649  0.22195620   0.35703797   0   0
#>   0.35703797  0.15981168   0.24746449   0   0
#>   0.39390279  0.18915837   0.44726911   0   0
#>   0.44726911  0.17041711   0.45484915   0   0
#>  -0.24734180  0.57247497   0.07261542   0   0
#>   0.07261542  0.52693664   0.12150988   0   0
#>   0.12150988  0.49990759   0.11713738   0   0
#>   0.11713738  0.49629811   0.11276370   0   0
#>   0.11276370  0.47404349   0.06276240   0   0
#>   0.06276240  0.47468362   0.03732853   0   0
#>   0.03732853  0.70824798  -0.64203604   0   0
#>  -0.64203604  0.72592844  -0.24734180   0   0
#>   0.34097316  0.38355021   0.35160141   0   0
#>   0.35160141  0.35416942   0.36900926   0   0
#>   0.24277723  0.52447888   0.30012838   0   0
#>   0.30012838  0.43414051   0.26878733   0   0
#>   0.26878733  0.42364933   0.23830134   0   0
#>   0.23830134  0.42600539   0.23215864   0   0
#>   0.23215864  0.52062255   0.18791009   0   0
#>   0.18791009  0.54070904   0.24277723   0   0
#>   0.05561421  0.70962402  -1.34896692   0   0
#>   0.16876885  0.91095425   0.36210103   0   0
#>   0.36210103  0.91008638   0.36812923   0   0
#>   0.36812923  0.90110813   0.38910405   0   0
#>   0.34564083  0.82415707   0.15231220   0   0
#>   0.15231220  0.86431212   0.16876885   0   0
#>   0.51836941  0.99130486   0.45035076   0   0
#>   0.45035076  1.17563981   0.40576800   0   0
#>   0.40576800  2.67479779   0.17757279   0   0
#>   0.17757279  4.77709136   0.06074523   0   1
#>   0.53691862  0.07976047   0.52038857   0   0
#>   0.81608837  0.16836453   0.79519686   0   0
#>   0.79519686  0.11087142   0.71687859   0   0
#>   0.83915894  0.12816112   0.81451041   0   0
#>   0.86890604 -0.17020229   0.76687936   0   0
#>   0.76687936 -0.14355555   0.75712888   0   0
#>   0.75712888 -0.07173445   0.74512029   0   0
#>   0.69813433  0.73543285   0.75193230   0   0
#>   0.75193230  0.73485570   0.75249730   0   0
#>   0.75249730  0.65374104   0.72158052   0   0
#>   0.72158052  0.64484067   0.61542627   0   0
#>   0.61542627  0.76845535   0.69813433   0   0
#>   0.58437048  0.71160439   0.45329623   0   0
#>   0.74003323  0.12900944   0.59081583   0   0
#>   0.08065540  0.16581257   0.09062069   0   0
#>   0.09062069  0.11854737   0.03459325   0   0
#>   0.03499327  0.04490232  -0.03949094   0   0
#>  -0.03949094 -4.25610113  -7.75569030   0   1
#>   0.59682648  0.48074425   0.62564947   0   0
#>   0.62564947  0.42958145   0.65036856   0   0
#>   0.65036856  0.39635454   0.61676322   0   0
#>   0.61676322  0.40052764   0.59775278   0   0
#>   0.59775278  0.44666008   0.52754033   0   0
#>   0.52754033  0.45580446   0.52775971   0   0
#>   0.52775971  0.49996871   0.59682648   0   0
#>   0.55187416  0.31998032   0.53090895   0   0
#>   0.63209137  0.31430354   0.59330196   0   0
#>   0.59330196  0.29879498   0.55187416   0   0
#>   0.48643203  0.44109400   0.42944135   0   0
#>   0.42944135  0.44417392   0.40156476   0   0
#>   0.40156476  0.58234680   0.36547015   0   0
#>   0.36547015  0.59273590   0.37061374   0   0
#>   0.37061374  0.61468774   0.44185728   0   0
#>   0.44185728  0.59703205   0.48634343   0   0
#>   0.48634343  0.59665486   0.48643203   0   0
#>   1.28660711  0.42355307   0.87190386   0   0
#>   0.87190386  0.44413515   0.84931035   0   0
#>   0.84931035  0.49924148   0.85260623   0   0
#>   0.85260623  0.51788858   0.87105701   0   0
#>   0.87105701  0.30297581   1.28660711   0   0
#>   0.44423409  0.75368870   0.34878622   0   0
#>   0.90032848 -7.34584652   8.51528784   0   1
#>   0.95629532  0.72995267   0.77279674   0   0
#>   0.77279674  0.73238667   0.76656816   0   0
#>   0.20140452  0.61077282   0.19444460   0   0
#>   0.45517663  0.24752947   0.46667353   0   0
#>   0.46667353  0.21878635   0.55997718   0   0
#>   0.55997718  0.13748579   0.54775774   0   0
#>   1.38366771  0.26698719   0.95176335   0   0
#>   0.07450151  0.36632831   0.07895623   0   0
#>   0.07895623  0.26358107   0.03908233   0   0
#>   0.61227716  0.51667714   0.63867400   0   0
#>   0.72190765  0.52416912   0.77791772   0   0
#>   0.77791772  0.44613141   0.67034458   0   0
#>   0.67034458  0.51117920   0.66269606   0   0
#>   0.66269606  0.56942748   0.72190765   0   0
#>   0.51236890  0.36617220   0.47546374   0   0
#>   0.32124297  0.69114876   0.21104313   0   0
#>   0.87646454  0.61883636   0.88952856   0   0
#>   0.88952856  0.53482516   0.88139215   0   0
#>   0.14762400  0.06072367   0.17567776   0   0
#>   0.55760569  0.58425000   0.60059165   0   0
#>   0.51823379  0.91076485   0.59794243   0   0
#>   0.23677505 -2.71822386  -0.01315880   0   1
#>   1.43040926  0.62466257   7.89495718   0   1
#>   0.78201702  3.48418036   1.10487630   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.908067292 0.889899719
#>   [2,] 0.322343710 0.166631506
#>   [3,] 0.233186218 0.674047236
#>   [4,] 0.640314733 0.998616298
#>   [5,] 0.873871870 0.603048689
#>   [6,] 0.091911938 0.415678295
#>   [7,] 0.800626542 0.506171667
#>   [8,] 0.298827561 0.747893915
#>   [9,] 0.157334033 0.001439569
#>  [10,] 0.254234404 0.321960222
#>  [11,] 0.157053021 0.411095057
#>  [12,] 0.504380923 0.055293699
#>  [13,] 0.322223036 0.303261384
#>  [14,] 0.505754792 0.222357066
#>  [15,] 0.440217475 0.033234417
#>  [16,] 0.847814258 0.341312416
#>  [17,] 1.000000000 0.508479259
#>  [18,] 0.125128759 0.461328722
#>  [19,] 0.173870063 0.763321330
#>  [20,] 0.151682555 0.822910980
#>  [21,] 0.119922538 0.838089070
#>  [22,] 0.668075771 0.651522026
#>  [23,] 0.735031567 0.524636989
#>  [24,] 0.855456930 0.766541144
#>  [25,] 0.114964496 0.638718470
#>  [26,] 0.484654057 0.177237135
#>  [27,] 0.148787045 0.055937952
#>  [28,] 0.131713276 0.000000000
#>  [29,] 0.476508738 0.605737730
#>  [30,] 0.328188098 0.573178890
#>  [31,] 0.484802776 0.438367039
#>  [32,] 0.973647643 0.034272737
#>  [33,] 0.431300786 0.971898452
#>  [34,] 0.737342499 0.382756278
#>  [35,] 0.103926275 0.885609174
#>  [36,] 0.805190030 0.856631952
#>  [37,] 0.158308574 0.847189367
#>  [38,] 0.373265849 0.171540335
#>  [39,] 0.594586989 0.139307998
#>  [40,] 0.711616700 0.528385086
#>  [41,] 0.480224336 0.450864277
#>  [42,] 0.349119741 0.276260084
#>  [43,] 0.191228118 0.495591275
#>  [44,] 0.363397828 0.952155291
#>  [45,] 0.333155970 0.007990054
#>  [46,] 0.500755000 0.621909591
#>  [47,] 0.499140661 0.687518906
#>  [48,] 0.977394259 0.359967915
#>  [49,] 0.363038885 0.531183428
#>  [50,] 0.173518891 0.613417420
#>  [51,] 0.731173977 0.318549644
#>  [52,] 0.277798421 0.522260122
#>  [53,] 0.204579468 0.401426928
#>  [54,] 0.493711121 0.641342690
#>  [55,] 0.444120176 0.053777016
#>  [56,] 0.617807994 0.783406831
#>  [57,] 0.446003844 0.726066401
#>  [58,] 0.136303890 0.790898493
#>  [59,] 0.070676366 0.052331581
#>  [60,] 0.972475208 0.394673374
#>  [61,] 0.686387284 0.971380579
#>  [62,] 0.517900567 0.579270020
#>  [63,] 0.438338360 0.109797000
#>  [64,] 0.957669620 0.314809390
#>  [65,] 0.884836496 0.596521865
#>  [66,] 0.170577054 0.951261934
#>  [67,] 0.079254404 0.594860651
#>  [68,] 0.536196151 0.651066264
#>  [69,] 0.297672549 0.611222038
#>  [70,] 0.884858146 0.599714490
#>  [71,] 0.932608006 0.453384951
#>  [72,] 0.235056049 0.109748577
#>  [73,] 0.290617834 0.587243375
#>  [74,] 0.687070928 0.623132046
#>  [75,] 0.032836275 0.055113963
#>  [76,] 0.384422070 0.696787419
#>  [77,] 0.932233908 0.583334996
#>  [78,] 0.834387504 0.821410167
#>  [79,] 0.134493312 0.067996260
#>  [80,] 0.310795500 0.468541140
#>  [81,] 0.281446633 0.361609877
#>  [82,] 0.323271266 0.873486489
#>  [83,] 0.597462516 0.993480659
#>  [84,] 0.720084143 0.365871210
#>  [85,] 0.581180342 0.243702207
#>  [86,] 0.923201170 0.932654111
#>  [87,] 0.627940521 0.654887113
#>  [88,] 0.293968755 0.108967836
#>  [89,] 0.461612678 0.349593602
#>  [90,] 0.396490928 0.428609974
#>  [91,] 0.174981115 0.128153323
#>  [92,] 0.754531530 0.882728004
#>  [93,] 0.448189122 0.080303589
#>  [94,] 0.554216826 0.204615443
#>  [95,] 0.485637254 1.000000000
#>  [96,] 0.933711302 0.956724087
#>  [97,] 0.000000000 0.416087142
#>  [98,] 0.175284317 0.366736710
#>  [99,] 0.005444609 0.525236220
#> [100,] 0.628609022 0.114874267
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.9080673,0.8898997) [6]: 17 24 36 77 78 86 
#> 2: (0.3223437,0.1666315) [7]: 10 13 38 42 72 88 91 
#> 3: (0.2331862,0.6740472) [6]: 8 19 25 50 69 73 
#> 4: (0.6403147,0.9986163) [4]: 61 83 95 96 
#> 5: (0.8738719,0.6030487) [6]: 7 23 24 65 70 74 
#> 6: (0.09191194,0.4156783) [6]: 11 18 91 97 98 99 
#> 7: (0.8006265,0.5061717) [6]: 5 16 23 34 65 71 
#> 8: (0.2988276,0.7478939) [7]: 3 19 37 57 69 76 82 
#> 9: (0.157334,0.001439569) [4]: 27 28 45 72 
#> 10: (0.2542344,0.3219602) [6]: 2 13 53 81 91 98 
#> 11: (0.157053,0.4110951) [5]: 6 18 43 53 98 
#> 12: (0.5043809,0.0552937) [8]: 15 26 32 39 55 63 93 100 
#> 13: (0.322223,0.3032614) [6]: 2 10 42 81 89 90 
#> 14: (0.5057548,0.2223571) [6]: 26 38 42 85 89 94 
#> 15: (0.4402175,0.03323442) [4]: 12 32 45 55 
#> 16: (0.8478143,0.3413124) [8]: 7 32 34 48 51 60 64 71 
#> 17: (1,0.5084793) [8]: 1 32 48 60 71 77 86 96 
#> 18: (0.1251288,0.4613287) [5]: 6 11 43 67 99 
#> 19: (0.1738701,0.7633213) [6]: 3 8 20 25 37 58 
#> 20: (0.1516826,0.822911) [4]: 19 21 37 58 
#> 21: (0.1199225,0.8380891) [6]: 20 35 37 58 67 99 
#> 22: (0.6680758,0.651522) [5]: 24 56 74 78 87 
#> 23: (0.7350316,0.524637) [5]: 5 7 34 40 74 
#> 24: (0.8554569,0.7665411) [7]: 1 5 22 70 74 77 78 
#> 25: (0.1149645,0.6387185) [5]: 3 19 50 58 67 
#> 26: (0.4846541,0.1772371) [6]: 12 14 38 39 63 94 
#> 27: (0.148787,0.05593795) [5]: 9 28 72 79 91 
#> 28: (0.1317133,0) [6]: 9 27 45 59 75 79 
#> 29: (0.4765087,0.6057377) [7]: 30 41 46 49 54 62 76 
#> 30: (0.3281881,0.5731789) [6]: 29 49 52 69 73 76 
#> 31: (0.4848028,0.438367) [7]: 40 41 62 84 85 89 90 
#> 32: (0.9736476,0.03427274) [9]: 12 15 16 17 45 48 51 64 100 
#> 33: (0.4313008,0.9718985) [5]: 44 56 57 82 95 
#> 34: (0.7373425,0.3827563) [6]: 7 16 23 40 51 84 
#> 35: (0.1039263,0.8856092) [4]: 21 37 66 99 
#> 36: (0.80519,0.856632) [5]: 1 56 78 86 92 
#> 37: (0.1583086,0.8471894) [7]: 8 19 20 21 35 66 82 
#> 38: (0.3732658,0.1715403) [6]: 2 14 26 42 63 88 
#> 39: (0.594587,0.139308) [5]: 12 26 85 94 100 
#> 40: (0.7116167,0.5283851) [6]: 23 31 34 62 74 84 
#> 41: (0.4802243,0.4508643) [5]: 29 31 49 62 90 
#> 42: (0.3491197,0.2762601) [5]: 2 13 14 38 89 
#> 43: (0.1912281,0.4955913) [7]: 11 18 50 52 53 67 80 
#> 44: (0.3633978,0.9521553) [4]: 33 66 82 95 
#> 45: (0.333156,0.007990054) [8]: 9 15 28 32 55 63 72 88 
#> 46: (0.500755,0.6219096) [4]: 29 54 62 68 
#> 47: (0.4991407,0.6875189) [4]: 54 56 57 68 
#> 48: (0.9773943,0.3599679) [5]: 16 17 32 60 64 
#> 49: (0.3630389,0.5311834) [6]: 29 30 41 52 80 90 
#> 50: (0.1735189,0.6134174) [6]: 3 25 43 52 67 73 
#> 51: (0.731174,0.3185496) [6]: 16 32 34 84 85 100 
#> 52: (0.2777984,0.5222601) [6]: 30 43 49 50 73 80 
#> 53: (0.2045795,0.4014269) [6]: 10 11 43 80 81 98 
#> 54: (0.4937111,0.6413427) [6]: 29 46 47 57 68 76 
#> 55: (0.4441202,0.05377702) [5]: 12 15 45 63 93 
#> 56: (0.617808,0.7834068) [12]: 22 33 36 47 57 61 68 78 83 87 92 95 
#> 57: (0.4460038,0.7260664) [7]: 8 33 47 54 56 76 82 
#> 58: (0.1363039,0.7908985) [5]: 19 20 21 25 67 
#> 59: (0.07067637,0.05233158) [4]: 28 75 79 91 
#> 60: (0.9724752,0.3946734) [4]: 16 17 48 71 
#> 61: (0.6863873,0.9713806) [5]: 4 56 83 92 96 
#> 62: (0.5179006,0.57927) [8]: 29 31 40 41 46 68 74 87 
#> 63: (0.4383384,0.109797) [7]: 12 26 38 45 55 88 93 
#> 64: (0.9576696,0.3148094) [3]: 16 32 48 
#> 65: (0.8848365,0.5965219) [5]: 5 7 70 71 77 
#> 66: (0.1705771,0.9512619) [5]: 35 37 44 82 95 
#> 67: (0.0792544,0.5948607) [7]: 18 21 25 43 50 58 99 
#> 68: (0.5361962,0.6510663) [6]: 46 47 54 56 62 87 
#> 69: (0.2976725,0.611222) [5]: 3 8 30 73 76 
#> 70: (0.8848581,0.5997145) [4]: 5 24 65 77 
#> 71: (0.932608,0.453385) [6]: 7 16 17 60 65 77 
#> 72: (0.235056,0.1097486) [6]: 2 9 27 45 88 91 
#> 73: (0.2906178,0.5872434) [5]: 3 30 50 52 69 
#> 74: (0.6870709,0.623132) [7]: 5 22 23 24 40 62 87 
#> 75: (0.03283627,0.05511396) [4]: 28 59 91 97 
#> 76: (0.3844221,0.6967874) [6]: 8 29 30 54 57 69 
#> 77: (0.9322339,0.583335) [6]: 1 17 24 65 70 71 
#> 78: (0.8343875,0.8214102) [5]: 1 22 24 36 56 
#> 79: (0.1344933,0.06799626) [4]: 27 28 59 91 
#> 80: (0.3107955,0.4685411) [6]: 43 49 52 53 81 90 
#> 81: (0.2814466,0.3616099) [5]: 10 13 53 80 90 
#> 82: (0.3232713,0.8734865) [6]: 8 33 37 44 57 66 
#> 83: (0.5974625,0.9934807) [4]: 4 56 61 95 
#> 84: (0.7200841,0.3658712) [5]: 31 34 40 51 85 
#> 85: (0.5811803,0.2437022) [8]: 14 31 39 51 84 89 94 100 
#> 86: (0.9232012,0.9326541) [5]: 1 17 36 92 96 
#> 87: (0.6279405,0.6548871) [5]: 22 56 62 68 74 
#> 88: (0.2939688,0.1089678) [5]: 2 38 45 63 72 
#> 89: (0.4616127,0.3495936) [6]: 13 14 31 42 85 90 
#> 90: (0.3964909,0.42861) [7]: 13 31 41 49 80 81 89 
#> 91: (0.1749811,0.1281533) [10]: 2 6 10 27 59 72 75 79 97 98 
#> 92: (0.7545315,0.882728) [5]: 36 56 61 86 96 
#> 93: (0.4481891,0.08030359) [3]: 12 55 63 
#> 94: (0.5542168,0.2046154) [4]: 14 26 39 85 
#> 95: (0.4856373,1) [6]: 4 33 44 56 66 83 
#> 96: (0.9337113,0.9567241) [5]: 4 17 61 86 92 
#> 97: (0,0.4160871) [4]: 6 75 91 99 
#> 98: (0.1752843,0.3667367) [5]: 6 10 11 53 91 
#> 99: (0.005444609,0.5252362) [6]: 6 18 21 35 67 97 
#> 100: (0.628609,0.1148743) [5]: 12 32 39 51 85 
#> number of nodes: 100 
#> number of arcs: 285 
#> number of boundary nodes: 12 
#> boundary nodes:  4 17 28 32 35 45 66 75 95 96 97 99 
#> number of triangles: 186 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.140378054 0.045750045 0.058775253 0.198643391 0.100973386 0.431071184 
#>                                                                         
#> 0.062830982 0.058917879 0.101011617 0.107467484 0.038111717 0.070513123 
#>                                                                         
#> 0.209385286 0.137825028 0.063786835 0.119988512 0.024994905 0.136676757 
#>                                                                         
#> 0.125906241 0.112015138 0.091629842 0.247757911 0.003192698 0.169397650 
#>                                                                         
#> 0.221178694 0.109551354 0.068144471 0.123509007 0.109284789 0.336933636 
#>                                                                         
#> 0.238583580 0.047958764 0.059519574 0.135677795 0.141899530 0.117989844 
#>                                                                         
#> 0.140535521 0.150898288 0.068187782 0.191822752 0.167045588 0.085299518 
#>                                                                         
#> 0.121847912 0.141344939 0.101675584 0.058485622 0.201601091 0.071184701 
#>                                                                         
#> 0.086567653 0.045405033 0.095106166 0.074451648 0.354327894 0.041886823 
#>                                                                         
#> 0.116292170 0.081812685 0.031094997 0.026836829 0.020910032 0.533431178 
#>                                                                         
#> 0.134285018 0.102393503 0.133132276 0.047484800 0.159716667 0.111533791 
#>                                                                         
#> 0.074756609 0.641030702 0.120039580 0.280991300 0.049278329 0.035052331 
#>                                                                         
#> 0.070967902 0.064502268 0.373639984 0.026264550 0.129950584 0.325716726 
#>                                                                         
#> 0.149641310 0.101466484 0.025166331 0.035514823 0.153668895 0.039450009 
#>                                                                         
#> 0.049952970 0.066584653 0.373587077 0.204170222 0.219888467 0.128918524 
#>                                                                         
#> 0.067117728 0.097874826 0.147883624 0.050127344 0.096073655 0.089703239 
#>                                                                         
#> 0.076777682 0.072512994 0.068053068 0.065711375 0.037942248 0.045957628 
#>                                                                         
#> 0.020670306 0.122548825 0.135798851 0.054572853 0.142068983 0.133819032 
#>                                                                         
#> 0.066235645 0.085706255 0.086640276 0.184985220 0.162734315 0.200287835 
#>                                                                         
#> 0.228042165 0.131269164 0.088311502 0.181120377 0.253735503 0.048603653 
#>                                                                         
#> 0.104793199 0.175903564 0.168991185 0.171360854 0.144371987 0.137281163 
#>                                                                         
#> 0.174764094 0.107890465 0.125642998 0.063043924 0.138505860 0.318807667 
#>                                                                         
#> 0.056317564 0.074090682 0.043583548 0.155481570 0.097232152 0.094542077 
#>                                                                         
#> 0.110885756 0.111816444 0.091823898 0.159812089 0.193857708 0.108122502 
#>                                                                         
#> 0.100595874 0.134108867 0.307515779 0.392343205 0.045353857 0.104186388 
#>                                                                         
#> 0.064266919 0.051158194 0.112851145 0.136629932 0.169605093 0.152303332 
#>                                                                         
#> 0.123387590 0.085065648 0.104083144 0.090030508 0.098803432 0.107183348 
#>                                                                         
#> 0.043158861 0.053520691 0.296372239 0.011481077 0.164526268 0.187877442 
#>                                                                         
#> 0.159452264 0.121449724 0.012760190 0.091912848 0.299284310 0.096675914 
#>                                                                         
#> 0.065302119 0.056456343 0.139569116 0.138694731 0.171479599 0.142146207 
#>                                                                         
#> 0.148786084 0.127949167 0.172061537 0.099690992 0.133309988 0.055164524 
#>                                                                         
#> 0.025661168 0.175943919 0.048089506 0.093704696 0.090763705 0.048499854 
#>                                                                         
#> 0.091145752 0.137776865 0.123270178 0.123528745 0.085628441 0.061506204 
#>                                                                         
#> 0.060279831 0.067849540 0.469737304 0.146888212 0.145698280 0.078387682 
#>                                                                         
#> 0.134676139 0.165650459 0.141900177 0.049810132 0.051607499 0.109997471 
#>                                                                         
#> 0.113007127 0.130916020 0.135601417 0.118840645 0.331824359 0.087046330 
#>                                                                         
#> 0.117087120 0.150221965 0.474938177 0.141192183 0.063586256 0.046601688 
#>                                                                         
#> 0.035200470 0.050140210 0.333139764 0.246604868 0.219866087 0.237742662 
#>                                                                         
#> 0.141139829 0.040276074 0.034158556 0.023712954 0.056557218 0.076819198 
#>                                                                         
#> 0.018700631 0.080399619 0.113199869 0.029144645 0.039542826 0.129499035 
#>                                                                         
#> 0.151852199 0.135770952 0.154918018 0.049130675 0.071636865 0.048769661 
#>                                                                         
#> 0.040116511 0.013309510 0.088849213 0.091752404 0.217216556 0.246197009 
#>                                                                         
#> 0.244024190 0.144738087 0.070714950 0.146134499 0.246271349 0.265167928 
#>                                                                         
#> 0.061173104 0.024144489 0.093555391 0.056984977 0.201181375 0.105251551 
#>                                                                         
#> 0.122589023 0.090585014 0.119149558 0.192822843 0.146383023 0.108315052 
#>                                                                         
#> 0.045893210 0.152566150 0.065646254 0.046494332 0.051979846 0.081568545 
#>                                                                         
#> 0.167631231 0.200093100 0.211056750 0.128951220 0.133516802 0.049197644 
#>                                     
#> 0.362463594 0.154683669 0.453119871 
#> 
#> $alpha
#> [1] 0.2828771
#> 
#> attr(,"class")
#> [1] "scree"