Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    71    9 0.075636410 0.77649116 0.075269165 0.83296727  0.075513271
#>    22   46 0.048795132 0.75691452 0.075266922 0.76712341  0.057990540
#>     1   22 0.049530773 0.81579012 0.048795132 0.75691452  0.047872695
#>    71   46 0.075636410 0.77649116 0.075266922 0.76712341  0.117693821
#>    24   39 0.190262466 0.74986142 0.159900956 0.77540483  0.124331107
#>     9   89 0.075269165 0.83296727 0.116049843 0.88878358  0.075119270
#>     9   39 0.075269165 0.83296727 0.159900956 0.77540483  0.136064032
#>     4    1 0.010247754 0.88814738 0.049530773 0.81579012  0.040237638
#>     4    9 0.010247754 0.88814738 0.075269165 0.83296727  0.055713149
#>   100   40 0.194908773 0.84693716 0.086182026 0.93148368  0.223242116
#>    89   39 0.116049843 0.88878358 0.159900956 0.77540483  0.136119026
#>   100   39 0.194908773 0.84693716 0.159900956 0.77540483  0.204684899
#>    40   64 0.086182026 0.93148368 0.079621335 0.94682442  0.070640830
#>    40   65 0.086182026 0.93148368 0.075003243 0.91879802  0.099168557
#>    18    4 0.002611201 0.98166772 0.010247754 0.88814738  0.026953229
#>     7   60 0.212798978 0.70571546 0.234162823 0.69509868  0.237839060
#>     7   24 0.212798978 0.70571546 0.190262466 0.74986142  0.122066260
#>    64   18 0.079621335 0.94682442 0.002611201 0.98166772  0.030391632
#>    28  100 0.272207059 0.72786522 0.194908773 0.84693716  0.246522893
#>     7   41 0.212798978 0.70571546 0.000000000 0.59506032  0.117627777
#>     7   46 0.212798978 0.70571546 0.075266922 0.76712341  0.105980072
#>    22    4 0.048795132 0.75691452 0.010247754 0.88814738 -0.219279200
#>    52   64 0.403198383 0.98868328 0.079621335 0.94682442  0.235857599
#>    13   52 0.314691663 0.84780915 0.403198383 0.98868328  0.253803395
#>    80   13 0.330093539 0.80781936 0.314691663 0.84780915  0.255131189
#>    28   80 0.272207059 0.72786522 0.330093539 0.80781936  0.254952019
#>    28   60 0.272207059 0.72786522 0.234162823 0.69509868  0.275921646
#>     7   14 0.212798978 0.70571546 0.042651864 0.53208440  0.142727690
#>    80   85 0.330093539 0.80781936 0.317304849 0.69655226  0.349494672
#>    50   43 0.235329475 0.52270911 0.263052958 0.47476562  0.274280708
#>    85   60 0.317304849 0.69655226 0.234162823 0.69509868  0.277248458
#>    52   18 0.403198383 0.98868328 0.002611201 0.98166772  0.197228627
#>    13   15 0.314691663 0.84780915 0.377904238 0.80402126  0.394696950
#>    45   53 0.371027391 0.44500687 0.426619725 0.46591104  0.376851579
#>    26   43 0.027739462 0.38671511 0.263052958 0.47476562  0.142942243
#>    85   50 0.317304849 0.69655226 0.235329475 0.52270911  0.304488973
#>    14   26 0.042651864 0.53208440 0.027739462 0.38671511  0.135182995
#>    14   41 0.042651864 0.53208440 0.000000000 0.59506032 -0.110428993
#>    58   43 0.310216561 0.49548667 0.263052958 0.47476562  0.301242996
#>    53   85 0.426619725 0.46591104 0.317304849 0.69655226  0.396973364
#>    53   29 0.426619725 0.46591104 0.554408333 0.53372814  0.432165859
#>    58    3 0.310216561 0.49548667 0.274476843 0.41629537  0.318796158
#>    27   52 0.546898624 0.83392038 0.403198383 0.98868328  0.450155271
#>    29   15 0.554408333 0.53372814 0.377904238 0.80402126  0.478123996
#>    45    3 0.371027391 0.44500687 0.274476843 0.41629537  0.327854731
#>     3   92 0.274476843 0.41629537 0.223838214 0.28811269  0.300447033
#>    26   92 0.027739462 0.38671511 0.223838214 0.28811269  0.152437934
#>    27   15 0.546898624 0.83392038 0.377904238 0.80402126  0.486595581
#>     8   42 0.373529854 0.38572813 0.386905701 0.34968543  0.425620460
#>    53    8 0.426619725 0.46591104 0.373529854 0.38572813  0.442906200
#>     8    3 0.373529854 0.38572813 0.274476843 0.41629537  0.305125856
#>    75   36 0.653426394 0.78916481 0.649780327 0.93159394  0.620344139
#>    27   29 0.546898624 0.83392038 0.554408333 0.53372814  0.546460067
#>    75   29 0.653426394 0.78916481 0.554408333 0.53372814  0.550539887
#>    53   83 0.426619725 0.46591104 0.411936160 0.33428153  0.461251179
#>    25   83 0.273964649 0.24064730 0.411936160 0.33428153  0.354970031
#>    36   84 0.649780327 0.93159394 0.515593735 1.00000000  0.560612715
#>    75   87 0.653426394 0.78916481 0.671000261 0.77697457  0.672055450
#>    32   73 0.539049819 0.48700133 0.531935072 0.36116576  0.553686654
#>    92   25 0.223838214 0.28811269 0.273964649 0.24064730  0.303990543
#>    30   25 0.122546479 0.17721805 0.273964649 0.24064730  0.197685726
#>    61   36 0.667149701 0.92198914 0.649780327 0.93159394  0.621312038
#>    61   75 0.667149701 0.92198914 0.653426394 0.78916481  0.665342881
#>    57   86 0.726918065 0.76749811 0.692043617 0.79364989  0.697712564
#>    26   30 0.027739462 0.38671511 0.122546479 0.17721805  0.104615182
#>    61   86 0.667149701 0.92198914 0.692043617 0.79364989  0.697502231
#>    94   34 0.802501251 0.85214845 0.757941559 0.89215407  0.743091170
#>    19   29 0.672481376 0.45077678 0.554408333 0.53372814  0.611266677
#>    32   53 0.539049819 0.48700133 0.426619725 0.46591104  0.492199970
#>    56   73 0.508440938 0.30275076 0.531935072 0.36116576  0.471039013
#>    96   57 0.696177491 0.71094918 0.726918065 0.76749811  0.694898313
#>     2   51 0.578509191 0.32088025 0.618839595 0.41540737  0.584070496
#>    54   79 0.173868831 0.09108353 0.218915503 0.11158830  0.174361689
#>    49   96 0.718701823 0.63298995 0.696177491 0.71094918  0.601510061
#>    19   49 0.672481376 0.45077678 0.718701823 0.63298995  0.655463785
#>    17   90 0.561167207 0.24842524 0.574873736 0.30629757  0.542788489
#>    81   56 0.485989468 0.25327934 0.508440938 0.30275076  0.453445764
#>    37   61 0.686140739 0.95760533 0.667149701 0.92198914  0.716343661
#>    57   94 0.726918065 0.76749811 0.802501251 0.85214845  0.744767136
#>    51   19 0.618839595 0.41540737 0.672481376 0.45077678  0.610517111
#>    56    2 0.508440938 0.30275076 0.578509191 0.32088025  0.540362782
#>    69    2 0.678073720 0.37482884 0.578509191 0.32088025  0.648521800
#>    48   74 0.313734094 0.11571568 0.363854260 0.13859299  0.312755748
#>    69   23 0.678073720 0.37482884 0.727756869 0.35148794  0.727197534
#>    69   82 0.678073720 0.37482884 0.664221812 0.40787957  0.635700942
#>    68   54 0.085789697 0.13827246 0.173868831 0.09108353  0.136536077
#>    81   74 0.485989468 0.25327934 0.363854260 0.13859299  0.389508606
#>    81   83 0.485989468 0.25327934 0.411936160 0.33428153  0.386589660
#>    79   33 0.218915503 0.11158830 0.185092078 0.08639655  0.205961947
#>    54   33 0.173868831 0.09108353 0.185092078 0.08639655  0.190315614
#>    48   79 0.313734094 0.11571568 0.218915503 0.11158830  0.264486716
#>    77   17 0.536025582 0.23438053 0.561167207 0.24842524  0.531364586
#>    81   77 0.485989468 0.25327934 0.536025582 0.23438053  0.520011759
#>    55   21 0.176797747 0.07568457 0.261751025 0.09238932  0.239148423
#>    10   33 0.228060614 0.09414783 0.185092078 0.08639655  0.212673545
#>    79   10 0.218915503 0.11158830 0.228060614 0.09414783  0.246023674
#>    84   37 0.515593735 1.00000000 0.686140739 0.95760533  0.684092824
#>    31   37 0.794894107 0.97528109 0.686140739 0.95760533  0.743454762
#>    81   59 0.485989468 0.25327934 0.389734911 0.13519223  0.475217459
#>    59   74 0.389734911 0.13519223 0.363854260 0.13859299  0.372015264
#>    23   19 0.727756869 0.35148794 0.672481376 0.45077678  0.729587234
#>    17   72 0.561167207 0.24842524 0.608465875 0.25216914  0.582954253
#>     2   90 0.578509191 0.32088025 0.574873736 0.30629757  0.626870924
#>    77   38 0.536025582 0.23438053 0.573906669 0.17084562  0.566442822
#>    31   34 0.794894107 0.97528109 0.757941559 0.89215407  0.803376593
#>    97   78 0.798372272 0.69304486 0.853385944 0.56106049  0.869104886
#>    77    6 0.536025582 0.23438053 0.410952287 0.10058164  0.482271846
#>    21    6 0.261751025 0.09238932 0.410952287 0.10058164  0.342655462
#>    35    6 0.520599489 0.04888296 0.410952287 0.10058164  0.350921296
#>    91   78 0.896231027 0.41396057 0.853385944 0.56106049  0.789332904
#>    69   72 0.678073720 0.37482884 0.608465875 0.25216914  0.671941571
#>    38   93 0.573906669 0.17084562 0.640399694 0.18854592  0.599641480
#>    38   72 0.573906669 0.17084562 0.608465875 0.25216914  0.587772079
#>    94   97 0.802501251 0.85214845 0.798372272 0.69304486  0.822239714
#>    97   49 0.798372272 0.69304486 0.718701823 0.63298995  0.795265577
#>    94   70 0.802501251 0.85214845 0.891702534 0.72749596  0.857685727
#>    98   94 0.827926991 0.97007163 0.802501251 0.85214845  0.850610569
#>    93   62 0.640399694 0.18854592 0.812633260 0.30405341  0.725505864
#>    93   23 0.640399694 0.18854592 0.727756869 0.35148794  0.703077534
#>     5   93 0.609671062 0.01617668 0.640399694 0.18854592  0.627874169
#>    55   35 0.176797747 0.07568457 0.520599489 0.04888296  0.313169374
#>    62   91 0.812633260 0.30405341 0.896231027 0.41396057  0.807571683
#>    23   91 0.727756869 0.35148794 0.896231027 0.41396057  0.787192702
#>    35    5 0.520599489 0.04888296 0.609671062 0.01617668  0.587131642
#>    70   78 0.891702534 0.72749596 0.853385944 0.56106049  0.981587229
#>    98   11 0.827926991 0.97007163 0.905110331 0.85899197  0.918931100
#>    16   11 0.999110025 0.90384724 0.905110331 0.85899197  0.999084859
#>    63   16 0.993893382 0.47895549 0.999110025 0.90384724  1.152590846
#>    12   99 0.732951519 0.10943540 0.844715549 0.23351285  0.741973410
#>    99   62 0.844715549 0.23351285 0.812633260 0.30405341  0.738839045
#>    88   63 0.905488915 0.30181317 0.993893382 0.47895549  1.049558129
#>    20   62 0.903331920 0.37010787 0.812633260 0.30405341  0.859823929
#>    91   20 0.896231027 0.41396057 0.903331920 0.37010787  0.973353292
#>    66   88 0.914506663 0.15958585 0.905488915 0.30181317  0.916201699
#>    67   66 0.801410314 0.06397652 0.914506663 0.15958585  0.827714823
#>    12    5 0.732951519 0.10943540 0.609671062 0.01617668  0.713914021
#>    66   44 0.914506663 0.15958585 0.978291364 0.14587473  0.960109682
#>    76   95 0.884204664 0.00000000 0.941227897 0.05388592  0.872111566
#>    95   66 0.941227897 0.05388592 0.914506663 0.15958585  0.873782392
#>    67    5 0.801410314 0.06397652 0.609671062 0.01617668  0.740548267
#>    47   44 1.000000000 0.15544665 0.978291364 0.14587473  1.022721634
#>     1    9 0.049530773 0.81579012 0.075269165 0.83296727  0.075513271
#>    71   22 0.075636410 0.77649116 0.048795132 0.75691452  0.047872695
#>    39   71 0.159900956 0.77540483 0.075636410 0.77649116  0.118143309
#>    39   46 0.159900956 0.77540483 0.075266922 0.76712341  0.117693821
#>    65   89 0.075003243 0.91879802 0.116049843 0.88878358  0.075119270
#>    65    9 0.075003243 0.91879802 0.075269165 0.83296727  0.055713149
#>    89   40 0.116049843 0.88878358 0.086182026 0.93148368  0.223242116
#>    41   22 0.000000000 0.59506032 0.048795132 0.75691452  0.104525267
#>    65   64 0.075003243 0.91879802 0.079621335 0.94682442  0.070640830
#>    65    4 0.075003243 0.91879802 0.010247754 0.88814738  0.026953229
#>    64  100 0.079621335 0.94682442 0.194908773 0.84693716  0.227452987
#>    46   24 0.075266922 0.76712341 0.190262466 0.74986142  0.122066260
#>    65   18 0.075003243 0.91879802 0.002611201 0.98166772  0.030391632
#>    24  100 0.190262466 0.74986142 0.194908773 0.84693716  0.246522893
#>    28    7 0.272207059 0.72786522 0.212798978 0.70571546  0.232506664
#>    41   46 0.000000000 0.59506032 0.075266922 0.76712341  0.105980072
#>    41    4 0.000000000 0.59506032 0.010247754 0.88814738 -0.219279200
#>    41   18 0.000000000 0.59506032 0.002611201 0.98166772 -1.651492722
#>   100   52 0.194908773 0.84693716 0.403198383 0.98868328  0.253803395
#>   100   13 0.194908773 0.84693716 0.314691663 0.84780915  0.255131189
#>   100   80 0.194908773 0.84693716 0.330093539 0.80781936  0.254952019
#>    85   28 0.317304849 0.69655226 0.272207059 0.72786522  0.322601063
#>    50   14 0.235329475 0.52270911 0.042651864 0.53208440  0.142727690
#>    50    7 0.235329475 0.52270911 0.212798978 0.70571546  0.177816865
#>    15   80 0.377904238 0.80402126 0.330093539 0.80781936  0.356790595
#>    15   85 0.377904238 0.80402126 0.317304849 0.69655226  0.349494672
#>    50   60 0.235329475 0.52270911 0.234162823 0.69509868  0.277248458
#>    52   15 0.403198383 0.98868328 0.377904238 0.80402126  0.394696950
#>    58   53 0.310216561 0.49548667 0.426619725 0.46591104  0.376851579
#>    58   50 0.310216561 0.49548667 0.235329475 0.52270911  0.304488973
#>    50   26 0.235329475 0.52270911 0.027739462 0.38671511  0.135182995
#>    26   41 0.027739462 0.38671511 0.000000000 0.59506032 -0.110428993
#>     3   43 0.274476843 0.41629537 0.263052958 0.47476562  0.301242996
#>     3   26 0.274476843 0.41629537 0.027739462 0.38671511  0.148643777
#>    58   85 0.310216561 0.49548667 0.317304849 0.69655226  0.396973364
#>    85   29 0.317304849 0.69655226 0.554408333 0.53372814  0.432165859
#>    84   52 0.515593735 1.00000000 0.403198383 0.98868328  0.148703076
#>    84   27 0.515593735 1.00000000 0.546898624 0.83392038  0.468380456
#>     8   45 0.373529854 0.38572813 0.371027391 0.44500687  0.413248638
#>    42   92 0.386905701 0.34968543 0.223838214 0.28811269  0.300447033
#>    83   42 0.411936160 0.33428153 0.386905701 0.34968543  0.425620460
#>    83    8 0.411936160 0.33428153 0.373529854 0.38572813  0.442906200
#>    42    3 0.386905701 0.34968543 0.274476843 0.41629537  0.305125856
#>    27   36 0.546898624 0.83392038 0.649780327 0.93159394  0.620344139
#>    96   29 0.696177491 0.71094918 0.554408333 0.53372814  0.550539887
#>    73   83 0.531935072 0.36116576 0.411936160 0.33428153  0.461251179
#>    86   87 0.692043617 0.79364989 0.671000261 0.77697457  0.672055450
#>    51   73 0.618839595 0.41540737 0.531935072 0.36116576  0.553686654
#>    42   25 0.386905701 0.34968543 0.273964649 0.24064730  0.303990543
#>    86   75 0.692043617 0.79364989 0.653426394 0.78916481  0.665342881
#>    96   75 0.696177491 0.71094918 0.653426394 0.78916481  0.617642429
#>    92   30 0.223838214 0.28811269 0.122546479 0.17721805  0.104615182
#>    37   36 0.686140739 0.95760533 0.649780327 0.93159394  0.611859228
#>    34   86 0.757941559 0.89215407 0.692043617 0.79364989  0.697502231
#>    32   29 0.539049819 0.48700133 0.554408333 0.53372814  0.611266677
#>    73   53 0.531935072 0.36116576 0.426619725 0.46591104  0.492199970
#>    87   57 0.671000261 0.77697457 0.726918065 0.76749811  0.694898313
#>    97   96 0.798372272 0.69304486 0.696177491 0.71094918  0.750123028
#>    74   25 0.363854260 0.13859299 0.273964649 0.24064730  0.375515662
#>    30   79 0.122546479 0.17721805 0.218915503 0.11158830  0.174361689
#>    68   26 0.085789697 0.13827246 0.027739462 0.38671511  0.005799405
#>    29   49 0.554408333 0.53372814 0.718701823 0.63298995  0.655463785
#>    56   90 0.508440938 0.30275076 0.574873736 0.30629757  0.542788489
#>    83   56 0.411936160 0.33428153 0.508440938 0.30275076  0.453445764
#>    34   61 0.757941559 0.89215407 0.667149701 0.92198914  0.716343661
#>    86   94 0.692043617 0.79364989 0.802501251 0.85214845  0.744767136
#>    32   19 0.539049819 0.48700133 0.672481376 0.45077678  0.610517111
#>    82   51 0.664221812 0.40787957 0.618839595 0.41540737  0.645203472
#>    73    2 0.531935072 0.36116576 0.578509191 0.32088025  0.540362782
#>    72    2 0.608465875 0.25216914 0.578509191 0.32088025  0.648521800
#>    82   23 0.664221812 0.40787957 0.727756869 0.35148794  0.727197534
#>    30   54 0.122546479 0.17721805 0.173868831 0.09108353  0.136536077
#>    74   83 0.363854260 0.13859299 0.411936160 0.33428153  0.386589660
#>    21   79 0.261751025 0.09238932 0.218915503 0.11158830  0.264486716
#>    97   57 0.798372272 0.69304486 0.726918065 0.76749811  0.806580653
#>    78   19 0.853385944 0.56106049 0.672481376 0.45077678  0.749316643
#>    25   48 0.273964649 0.24064730 0.313734094 0.11571568  0.263930457
#>    56   17 0.508440938 0.30275076 0.561167207 0.24842524  0.531364586
#>    56   77 0.508440938 0.30275076 0.536025582 0.23438053  0.520011759
#>    10   21 0.228060614 0.09414783 0.261751025 0.09238932  0.239148423
#>    55   33 0.176797747 0.07568457 0.185092078 0.08639655  0.212673545
#>    55   54 0.176797747 0.07568457 0.173868831 0.09108353  0.177408557
#>    55   68 0.176797747 0.07568457 0.085789697 0.13827246  0.105997939
#>    51   69 0.618839595 0.41540737 0.678073720 0.37482884  0.622897792
#>    34   37 0.757941559 0.89215407 0.686140739 0.95760533  0.743454762
#>     6   59 0.410952287 0.10058164 0.389734911 0.13519223  0.475217459
#>     6   81 0.410952287 0.10058164 0.485989468 0.25327934  0.479904973
#>     6   74 0.410952287 0.10058164 0.363854260 0.13859299  0.372015264
#>     6   48 0.410952287 0.10058164 0.313734094 0.11571568  0.358560549
#>    82   19 0.664221812 0.40787957 0.672481376 0.45077678  0.729587234
#>    90   72 0.574873736 0.30629757 0.608465875 0.25216914  0.582954253
#>    17   38 0.561167207 0.24842524 0.573906669 0.17084562  0.566442822
#>    98   34 0.827926991 0.97007163 0.757941559 0.89215407  0.803376593
#>    38    6 0.573906669 0.17084562 0.410952287 0.10058164  0.482271846
#>    23   72 0.727756869 0.35148794 0.608465875 0.25216914  0.671941571
#>    72   93 0.608465875 0.25216914 0.640399694 0.18854592  0.599641480
#>    70   97 0.891702534 0.72749596 0.798372272 0.69304486  0.822239714
#>    78   49 0.853385944 0.56106049 0.718701823 0.63298995  0.795265577
#>    11   70 0.905110331 0.85899197 0.891702534 0.72749596  0.857685727
#>    11   94 0.905110331 0.85899197 0.802501251 0.85214845  0.850610569
#>    23   62 0.727756869 0.35148794 0.812633260 0.30405341  0.725505864
#>    38   35 0.573906669 0.17084562 0.520599489 0.04888296  0.493432166
#>    21   35 0.261751025 0.09238932 0.520599489 0.04888296  0.313169374
#>    19   91 0.672481376 0.45077678 0.896231027 0.41396057  0.787192702
#>    38    5 0.573906669 0.17084562 0.609671062 0.01617668  0.587131642
#>    63   78 0.993893382 0.47895549 0.853385944 0.56106049  0.981587229
#>    63   91 0.993893382 0.47895549 0.896231027 0.41396057  0.910770642
#>    70   16 0.891702534 0.72749596 0.999110025 0.90384724  1.152590846
#>    93   99 0.640399694 0.18854592 0.844715549 0.23351285  0.741973410
#>     5   55 0.609671062 0.01617668 0.176797747 0.07568457  0.287418343
#>    88   99 0.905488915 0.30181317 0.844715549 0.23351285  0.858565161
#>    20   63 0.903331920 0.37010787 0.993893382 0.47895549  1.049558129
#>    88   62 0.905488915 0.30181317 0.812633260 0.30405341  0.859823929
#>    99   66 0.844715549 0.23351285 0.914506663 0.15958585  0.827714823
#>    67   12 0.801410314 0.06397652 0.732951519 0.10943540  0.810506447
#>    47   66 1.000000000 0.15544665 0.914506663 0.15958585  0.960952953
#>    67   95 0.801410314 0.06397652 0.941227897 0.05388592  0.872111566
#>    47   63 1.000000000 0.15544665 0.993893382 0.47895549 13.008163931
#>    47   88 1.000000000 0.15544665 0.905488915 0.30181317  1.092711487
#>    76    5 0.884204664 0.00000000 0.609671062 0.01617668  0.740548267
#>    95   44 0.941227897 0.05388592 0.978291364 0.14587473  1.022721634
#>     1   71 0.049530773 0.81579012 0.075636410 0.77649116  0.075513271
#>    89  100 0.116049843 0.88878358 0.194908773 0.84693716  0.223242116
#>    24   28 0.190262466 0.74986142 0.272207059 0.72786522  0.246522893
#>    58   45 0.310216561 0.49548667 0.371027391 0.44500687  0.376851579
#>    84   18 0.515593735 1.00000000 0.002611201 0.98166772  0.148703076
#>    27   75 0.546898624 0.83392038 0.653426394 0.78916481  0.620344139
#>    51   32 0.618839595 0.41540737 0.539049819 0.48700133  0.553686654
#>    96   87 0.696177491 0.71094918 0.671000261 0.77697457  0.617642429
#>    79   25 0.218915503 0.11158830 0.273964649 0.24064730  0.204509662
#>    68   30 0.085789697 0.13827246 0.122546479 0.17721805  0.005799405
#>    21   48 0.261751025 0.09238932 0.313734094 0.11571568  0.264486716
#>    10   55 0.228060614 0.09414783 0.176797747 0.07568457  0.239148423
#>    31   84 0.794894107 0.97528109 0.515593735 1.00000000  0.684092824
#>    98   31 0.827926991 0.97007163 0.794894107 0.97528109  0.803376593
#>    63   70 0.993893382 0.47895549 0.891702534 0.72749596  0.981587229
#>    16   98 0.999110025 0.90384724 0.827926991 0.97007163  0.918931100
#>    12   93 0.732951519 0.10943540 0.640399694 0.18854592  0.643937843
#>    20   88 0.903331920 0.37010787 0.905488915 0.30181317  1.049558129
#>    99   67 0.844715549 0.23351285 0.801410314 0.06397652  0.827714823
#>    67   76 0.801410314 0.06397652 0.884204664 0.00000000  0.872111566
#>    47   16 1.000000000 0.15544665 0.999110025 0.90384724 13.008163931
#>    95   47 0.941227897 0.05388592 1.000000000 0.15544665  1.022721634
#>           my1          mx2          my2 bp1 bp2
#>   0.804729609  0.118143309  0.805006817   0   0
#>   0.772495996  0.104525267  0.651830850   0   0
#>   0.786368438 -0.087820133  0.788063897   0   0
#>   0.770141146  0.057990540  0.772495996   0   0
#>   0.702309733  0.204684899  0.797820183   0   0
#>   0.875882591  0.136064032  0.831354950   0   0
#>   0.831354950  0.118143309  0.805006817   0   0
#>   0.857586919 -0.087820133  0.788063897   0   0
#>   0.875822467  0.040237638  0.857586919   0   0
#>   0.995558320  0.227452987  1.000973497   0   0
#>   0.831376220  0.136064032  0.831354950   0   0
#>   0.797820183  0.136119026  0.831376220   0   0
#>   0.933910519  0.227452987  1.000973497   0   0
#>   0.908771478  0.070640830  0.933910519   0   0
#>   0.936583450 -1.651492722  0.799527253   0   0
#>   0.729299590  0.177816865  0.608518628   0   0
#>   0.687221828  0.232506664  0.743601659   0   0
#>   0.940542639  0.197228627  1.309283276   0   0
#>   0.795817703  0.254952019  0.801289665   0   0
#>   0.628794965  0.105980072  0.651194463   0   0
#>   0.651194463  0.122066260  0.687221828   0   0
#>   0.749450077 -0.087820133  0.788063897   0   0
#>   1.010673893  0.197228627  1.309283276   0   0
#>   0.984303346  0.394696950  0.895784422   0   0
#>   0.801908846  0.356790595  0.841062479   0   0
#>   0.801289665  0.322601063  0.752312025   0   0
#>   0.685083055  0.237839060  0.729299590   0   0
#>   0.604198687  0.117627777  0.628794965   0   0
#>   0.749220958  0.322601063  0.752312025   0   0
#>   0.513245448  0.142942243  0.437298550   0   0
#>   0.609191535  0.275921646  0.685083055   0   0
#>   1.309283276  0.148703076  4.080081587   0   0
#>   0.895784422  0.356790595  0.841062479   0   0
#>   0.513891002  0.413248638  0.417097053   0   0
#>   0.437298550  0.148643777  0.422061297   0   0
#>   0.596346326  0.277248458  0.609191535   0   0
#>   0.449142766 -0.110428993  0.474338354   0   0
#>   0.474338354  0.117627777  0.628794965   0   0
#>   0.451876047  0.274280708  0.513245448   0   0
#>   0.593085917  0.432165859  0.609765777   0   0
#>   0.609765777  0.471857656  0.534974028   0   0
#>   0.443954154  0.301242996  0.451876047   0   0
#>   0.888187999  0.468380456  0.905110424   0   0
#>   0.676689736  0.486595581  0.682221765   0   0
#>   0.413492134  0.318796158  0.443954154   0   0
#>   0.331942088  0.152437934  0.390413152   0   0
#>   0.390413152  0.104615182  0.295304107   0   0
#>   0.682221765  0.450155271  0.888187999   0   0
#>   0.384556218  0.305125856  0.339839327   0   0
#>   0.397460500  0.413248638  0.417097053   0   0
#>   0.339839327  0.327854731  0.413492134   0   0
#>   0.859579168  0.621312038  0.859603945   0   0
#>   0.683719355  0.486595581  0.682221765   0   0
#>   0.682137844  0.546460067  0.683719355   0   0
#>   0.395414076  0.442906200  0.397460500   0   0
#>   0.269753303  0.375515662  0.239478992   0   0
#>   0.922495582  0.611859228  1.023021696   0   0
#>   0.797258434  0.617642429  0.718814751   0   0
#>   0.423054846  0.492199970  0.426531305   0   0
#>   0.322557558  0.197685726  0.210292992   0   0
#>   0.210292992  0.204509662  0.194002904   0   0
#>   0.859603945  0.668159008  0.944322199   0   0
#>   0.855054713  0.621312038  0.859603945   0   0
#>   0.764880528  0.744767136  0.827629699   0   0
#>   0.295304107  0.005799405  0.250585440   0   0
#>   0.861292650  0.665342881  0.855054713   0   0
#>   0.830794276  0.815201064  0.911112851   0   0
#>   0.489152038  0.655463785  0.552062239   0   0
#>   0.426531305  0.471857656  0.534974028   0   0
#>   0.351725669  0.540362782  0.323844100   0   0
#>   0.748274467  0.750123028  0.718253749   0   0
#>   0.374374622  0.622897792  0.357808789   0   0
#>   0.149734400  0.190315614  0.114685420   0   0
#>   0.641363913  0.743857569  0.682491592   0   0
#>   0.552062239  0.749316643  0.528255392   0   0
#>   0.283337370  0.582954253  0.273824478   0   0
#>   0.297878810  0.520011759  0.267669359   0   0
#>   0.918629488  0.668159008  0.944322199   0   0
#>   0.827629699  0.806580653  0.772437210   0   0
#>   0.486391046  0.645203472  0.433785185   0   0
#>   0.323844100  0.540694254  0.322563000   0   0
#>   0.310518543  0.622897792  0.357808789   0   0
#>   0.184199951  0.358560549  0.083849693   0   0
#>   0.414845252  0.671941571  0.297228117   0   0
#>   0.376498068  0.727197534  0.414845252   0   0
#>   0.127196405  0.105997939  0.070196318   0   0
#>   0.233649509  0.386589660  0.236758041   0   0
#>   0.236758041  0.453445764  0.297878810   0   0
#>   0.093678050  0.190315614  0.114685420   0   0
#>   0.114685420  0.177408557  0.083778770   0   0
#>   0.155878412  0.263930457  0.168657410   0   0
#>   0.272249768  0.566442822  0.209455763   0   0
#>   0.267669359  0.479904973  0.161483461   0   0
#>  -0.017033977  0.313169374 -0.393473132   0   0
#>   0.056472946  0.205961947  0.093678050   0   0
#>   0.114684879  0.205961947  0.093678050   0   0
#>   1.313605968  0.611859228  1.023021696   0   0
#>   0.948370691  0.684092824  1.313605968   0   0
#>   0.163786952  0.389508606  0.233649509   0   0
#>   0.100520779  0.389508606  0.233649509   0   0
#>   0.417537668  0.787192702  0.449607443   0   0
#>   0.273824478  0.587772079  0.212958272   0   0
#>   0.301079190  0.540694254  0.322563000   0   0
#>   0.209455763  0.482271846  0.159270942   0   0
#>   0.921733573  0.743454762  0.948370691   0   0
#>   0.645070029  0.795265577  0.614292358   0   0
#>   0.159270942  0.479904973  0.161483461   0   0
#>  -0.018321539  0.350921296 -0.168861647   0   0
#>  -0.168861647  0.493432166  0.133388221   0   0
#>   0.462614473  0.910770642  0.497985056   0   0
#>   0.297228117  0.648521800  0.310518543   0   0
#>   0.207914256  0.627874169  0.101855224   0   0
#>   0.212958272  0.599641480  0.207914256   0   0
#>   0.772030834  0.806580653  0.772437210   0   0
#>   0.614292358  0.743857569  0.682491592   0   0
#>   0.797395989  0.822239714  0.772030834   0   0
#>   0.903478110  0.815201064  0.911112851   0   0
#>   0.247806594  0.738839045  0.227925448   0   0
#>   0.259830967  0.725505864  0.247806594   0   0
#>   0.101855224  0.643937843  0.098991518   0   0
#>  -0.393473132  0.287418343 -0.723798477   0   0
#>   0.394650074  0.826591855  0.380182919   0   0
#>   0.449607443  0.807571683  0.394650074   0   0
#>   0.092434252  0.287418343 -0.723798477   0   0
#>   0.619174467  0.869104886  0.645070029   0   0
#>   0.950950410  0.850610569  0.903478110   0   0
#>   0.782978435  0.918931100  0.950950410   0   0
#>   0.689484968 13.008163931  0.543927207   0   0
#>   0.213683860  0.810506447  0.151952030   0   0
#>   0.227925448  0.858565161  0.282377637   0   0
#>   0.340544816  1.092711487  0.319008744   0   0
#>   0.334552319  0.826591855  0.380182919   0   0
#>   0.403947411  0.826591855  0.380182919   0   0
#>   0.231092863  0.960952953  0.233930261   0   0
#>   0.147556437  0.873782392  0.093063077   0   0
#>   0.006488689  0.643937843  0.098991518   0   0
#>   0.216512868  0.937008857  0.109046902   0   0
#>   0.069911736  9.626446350 -9.194112208   0   1
#>   0.093063077  0.937008857  0.109046902   0   0
#>  -0.100349139  0.713914021  0.006488689   0   0
#>   0.074512126  0.960109682  0.216512868   0   0
#>   0.804729609  0.040237638  0.857586919   0   0
#>   0.786368438  0.057990540  0.772495996   0   0
#>   0.805006817  0.117693821  0.770141146   0   0
#>   0.770141146  0.124331107  0.702309733   0   0
#>   0.875882591  0.099168557  0.908771478   0   0
#>   0.875822467  0.075119270  0.875882591   0   0
#>   0.995558320  0.099168557  0.908771478   0   0
#>   0.651830850 -0.219279200  0.749450077   0   0
#>   0.933910519  0.030391632  0.940542639   0   0
#>   0.936583450  0.055713149  0.875822467   0   0
#>   1.000973497  0.235857599  1.010673893   0   0
#>   0.687221828  0.124331107  0.702309733   0   0
#>   0.940542639  0.026953229  0.936583450   0   0
#>   0.795817703  0.204684899  0.797820183   0   0
#>   0.743601659  0.237839060  0.729299590   0   0
#>   0.651194463  0.104525267  0.651830850   0   0
#>   0.749450077 -1.651492722  0.799527253   0   0
#>   0.799527253 -4.237978089  0.816996744   0   1
#>   0.984303346  0.235857599  1.010673893   0   0
#>   0.801908846  0.253803395  0.984303346   0   0
#>   0.801289665  0.255131189  0.801908846   0   0
#>   0.752312025  0.275921646  0.685083055   0   0
#>   0.604198687  0.135182995  0.449142766   0   0
#>   0.608518628  0.142727690  0.604198687   0   0
#>   0.841062479  0.349494672  0.749220958   0   0
#>   0.749220958  0.478123996  0.676689736   0   0
#>   0.609191535  0.177816865  0.608518628   0   0
#>   0.895784422  0.450155271  0.888187999   0   0
#>   0.513891002  0.396973364  0.593085917   0   0
#>   0.596346326  0.274280708  0.513245448   0   0
#>   0.449142766  0.142942243  0.437298550   0   0
#>   0.474338354 -4.826553860 -0.153575132   0   1
#>   0.451876047  0.148643777  0.422061297   0   0
#>   0.422061297  0.152437934  0.390413152   0   0
#>   0.593085917  0.304488973  0.596346326   0   0
#>   0.609765777  0.478123996  0.676689736   0   0
#>   4.080081587  0.468380456  0.905110424   0   0
#>   0.905110424  0.560612715  0.922495582   0   0
#>   0.417097053  0.327854731  0.413492134   0   0
#>   0.331942088  0.303990543  0.322557558   0   0
#>   0.384556218  0.354970031  0.269753303   0   0
#>   0.397460500  0.425620460  0.384556218   0   0
#>   0.339839327  0.300447033  0.331942088   0   0
#>   0.859579168  0.560612715  0.922495582   0   0
#>   0.682137844  0.601510061  0.641363913   0   0
#>   0.395414076  0.471039013  0.351725669   0   0
#>   0.797258434  0.697712564  0.764880528   0   0
#>   0.423054846  0.584070496  0.374374622   0   0
#>   0.322557558  0.354970031  0.269753303   0   0
#>   0.855054713  0.672055450  0.797258434   0   0
#>   0.718814751  0.550539887  0.682137844   0   0
#>   0.295304107  0.197685726  0.210292992   0   0
#>   1.023021696  0.668159008  0.944322199   0   0
#>   0.861292650  0.743091170  0.830794276   0   0
#>   0.489152038  0.471857656  0.534974028   0   0
#>   0.426531305  0.461251179  0.395414076   0   0
#>   0.748274467  0.697712564  0.764880528   0   0
#>   0.718253749  0.743857569  0.682491592   0   0
#>   0.239478992  0.312755748  0.184199951   0   0
#>   0.149734400  0.204509662  0.194002904   0   0
#>   0.250585440 -3.810900299 -0.641211177   0   1
#>   0.552062239  0.601510061  0.641363913   0   0
#>   0.283337370  0.540694254  0.322563000   0   0
#>   0.297878810  0.471039013  0.351725669   0   0
#>   0.918629488  0.697502231  0.861292650   0   0
#>   0.827629699  0.743091170  0.830794276   0   0
#>   0.486391046  0.611266677  0.489152038   0   0
#>   0.433785185  0.635700942  0.376498068   0   0
#>   0.323844100  0.584070496  0.374374622   0   0
#>   0.310518543  0.626870924  0.301079190   0   0
#>   0.414845252  0.729587234  0.417537668   0   0
#>   0.127196405  0.174361689  0.149734400   0   0
#>   0.236758041  0.375515662  0.239478992   0   0
#>   0.155878412  0.246023674  0.114684879   0   0
#>   0.772437210  0.750123028  0.718253749   0   0
#>   0.528255392  0.789332904  0.462614473   0   0
#>   0.168657410  0.312755748  0.184199951   0   0
#>   0.272249768  0.542788489  0.283337370   0   0
#>   0.267669359  0.531364586  0.272249768   0   0
#>  -0.017033977  0.246023674  0.114684879   0   0
#>   0.056472946  0.177408557  0.083778770   0   0
#>   0.083778770  0.105997939  0.070196318   0   0
#>   0.070196318 -5.024279342 -7.389657601   0   1
#>   0.357808789  0.635700942  0.376498068   0   0
#>   0.948370691  0.716343661  0.918629488   0   0
#>   0.163786952  0.372015264  0.100520779   0   0
#>   0.161483461  0.475217459  0.163786952   0   0
#>   0.100520779  0.358560549  0.083849693   0   0
#>   0.083849693  0.342655462 -0.018321539   0   0
#>   0.417537668  0.645203472  0.433785185   0   0
#>   0.273824478  0.626870924  0.301079190   0   0
#>   0.209455763  0.587772079  0.212958272   0   0
#>   0.921733573  0.815201064  0.911112851   0   0
#>   0.159270942  0.493432166  0.133388221   0   0
#>   0.297228117  0.703077534  0.259830967   0   0
#>   0.207914256  0.703077534  0.259830967   0   0
#>   0.772030834  0.869104886  0.645070029   0   0
#>   0.614292358  0.749316643  0.528255392   0   0
#>   0.797395989  0.999084859  0.782978435   0   0
#>   0.903478110  0.857685727  0.797395989   0   0
#>   0.247806594  0.807571683  0.394650074   0   0
#>   0.133388221  0.587131642  0.092434252   0   0
#>  -0.393473132  0.350921296 -0.168861647   0   0
#>   0.449607443  0.789332904  0.462614473   0   0
#>   0.092434252  0.627874169  0.101855224   0   0
#>   0.619174467  0.910770642  0.497985056   0   0
#>   0.497985056  0.973353292  0.403947411   0   0
#>   0.689484968  0.999084859  0.782978435   0   0
#>   0.213683860  0.738839045  0.227925448   0   0
#>  -0.723798477 -0.024270998 -2.991094693   0   1
#>   0.282377637  0.916201699  0.231092863   0   0
#>   0.340544816  0.973353292  0.403947411   0   0
#>   0.334552319  0.858565161  0.282377637   0   0
#>   0.147556437  0.916201699  0.231092863   0   0
#>   0.151952030  0.713914021  0.006488689   0   0
#>   0.233930261  0.960109682  0.216512868   0   0
#>   0.069911736  0.873782392  0.093063077   0   0
#>   0.543927207  1.092711487  0.319008744   0   0
#>   0.319008744  0.960952953  0.233930261   0   0
#>  -0.100349139  0.526656949 -3.730287115   0   1
#>   0.074512126  0.937008857  0.109046902   0   0
#>   0.804729609  0.047872695  0.786368438   0   0
#>   0.995558320  0.136119026  0.831376220   0   0
#>   0.795817703  0.232506664  0.743601659   0   0
#>   0.513891002  0.318796158  0.443954154   0   0
#>   4.080081587  0.079127457  6.026979296   0   1
#>   0.859579168  0.546460067  0.683719355   0   0
#>   0.423054846  0.610517111  0.486391046   0   0
#>   0.718814751  0.694898313  0.748274467   0   0
#>   0.194002904  0.263930457  0.168657410   0   0
#>   0.250585440  0.136536077  0.127196405   0   0
#>   0.155878412  0.342655462 -0.018321539   0   0
#>  -0.017033977  0.212673545  0.056472946   0   0
#>   1.313605968  0.998503859  4.866154422   0   1
#>   0.921733573  5.461715410 30.459951605   0   1
#>   0.619174467  1.152590846  0.689484968   0   0
#>   0.950950410  2.884670467  6.032178965   0   1
#>   0.098991518  0.741973410  0.213683860   0   0
#>   0.340544816  0.859823929  0.334552319   0   0
#>   0.147556437  0.810506447  0.151952030   0   0
#>   0.069911736  0.740548267 -0.100349139   0   0
#>   0.543927207 14.344344840  0.545516152   0   1
#>   0.074512126  8.398902572 -4.194004698   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.049530773 0.81579012
#>   [2,] 0.578509191 0.32088025
#>   [3,] 0.274476843 0.41629537
#>   [4,] 0.010247754 0.88814738
#>   [5,] 0.609671062 0.01617668
#>   [6,] 0.410952287 0.10058164
#>   [7,] 0.212798978 0.70571546
#>   [8,] 0.373529854 0.38572813
#>   [9,] 0.075269165 0.83296727
#>  [10,] 0.228060614 0.09414783
#>  [11,] 0.905110331 0.85899197
#>  [12,] 0.732951519 0.10943540
#>  [13,] 0.314691663 0.84780915
#>  [14,] 0.042651864 0.53208440
#>  [15,] 0.377904238 0.80402126
#>  [16,] 0.999110025 0.90384724
#>  [17,] 0.561167207 0.24842524
#>  [18,] 0.002611201 0.98166772
#>  [19,] 0.672481376 0.45077678
#>  [20,] 0.903331920 0.37010787
#>  [21,] 0.261751025 0.09238932
#>  [22,] 0.048795132 0.75691452
#>  [23,] 0.727756869 0.35148794
#>  [24,] 0.190262466 0.74986142
#>  [25,] 0.273964649 0.24064730
#>  [26,] 0.027739462 0.38671511
#>  [27,] 0.546898624 0.83392038
#>  [28,] 0.272207059 0.72786522
#>  [29,] 0.554408333 0.53372814
#>  [30,] 0.122546479 0.17721805
#>  [31,] 0.794894107 0.97528109
#>  [32,] 0.539049819 0.48700133
#>  [33,] 0.185092078 0.08639655
#>  [34,] 0.757941559 0.89215407
#>  [35,] 0.520599489 0.04888296
#>  [36,] 0.649780327 0.93159394
#>  [37,] 0.686140739 0.95760533
#>  [38,] 0.573906669 0.17084562
#>  [39,] 0.159900956 0.77540483
#>  [40,] 0.086182026 0.93148368
#>  [41,] 0.000000000 0.59506032
#>  [42,] 0.386905701 0.34968543
#>  [43,] 0.263052958 0.47476562
#>  [44,] 0.978291364 0.14587473
#>  [45,] 0.371027391 0.44500687
#>  [46,] 0.075266922 0.76712341
#>  [47,] 1.000000000 0.15544665
#>  [48,] 0.313734094 0.11571568
#>  [49,] 0.718701823 0.63298995
#>  [50,] 0.235329475 0.52270911
#>  [51,] 0.618839595 0.41540737
#>  [52,] 0.403198383 0.98868328
#>  [53,] 0.426619725 0.46591104
#>  [54,] 0.173868831 0.09108353
#>  [55,] 0.176797747 0.07568457
#>  [56,] 0.508440938 0.30275076
#>  [57,] 0.726918065 0.76749811
#>  [58,] 0.310216561 0.49548667
#>  [59,] 0.389734911 0.13519223
#>  [60,] 0.234162823 0.69509868
#>  [61,] 0.667149701 0.92198914
#>  [62,] 0.812633260 0.30405341
#>  [63,] 0.993893382 0.47895549
#>  [64,] 0.079621335 0.94682442
#>  [65,] 0.075003243 0.91879802
#>  [66,] 0.914506663 0.15958585
#>  [67,] 0.801410314 0.06397652
#>  [68,] 0.085789697 0.13827246
#>  [69,] 0.678073720 0.37482884
#>  [70,] 0.891702534 0.72749596
#>  [71,] 0.075636410 0.77649116
#>  [72,] 0.608465875 0.25216914
#>  [73,] 0.531935072 0.36116576
#>  [74,] 0.363854260 0.13859299
#>  [75,] 0.653426394 0.78916481
#>  [76,] 0.884204664 0.00000000
#>  [77,] 0.536025582 0.23438053
#>  [78,] 0.853385944 0.56106049
#>  [79,] 0.218915503 0.11158830
#>  [80,] 0.330093539 0.80781936
#>  [81,] 0.485989468 0.25327934
#>  [82,] 0.664221812 0.40787957
#>  [83,] 0.411936160 0.33428153
#>  [84,] 0.515593735 1.00000000
#>  [85,] 0.317304849 0.69655226
#>  [86,] 0.692043617 0.79364989
#>  [87,] 0.671000261 0.77697457
#>  [88,] 0.905488915 0.30181317
#>  [89,] 0.116049843 0.88878358
#>  [90,] 0.574873736 0.30629757
#>  [91,] 0.896231027 0.41396057
#>  [92,] 0.223838214 0.28811269
#>  [93,] 0.640399694 0.18854592
#>  [94,] 0.802501251 0.85214845
#>  [95,] 0.941227897 0.05388592
#>  [96,] 0.696177491 0.71094918
#>  [97,] 0.798372272 0.69304486
#>  [98,] 0.827926991 0.97007163
#>  [99,] 0.844715549 0.23351285
#> [100,] 0.194908773 0.84693716
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,71,9), [4,10,3]
#> 2: (71,22,46), [15,5,3]
#> 3: (71,1,22), [9,2,1]
#> 4: (39,9,71), [1,5,8]
#> 5: (39,71,46), [2,6,4]
#> 6: (46,24,39), [14,5,21]
#> 7: (65,9,89), [8,17,11]
#> 8: (89,9,39), [4,13,7]
#> 9: (4,22,1), [3,10,27]
#> 10: (9,4,1), [9,1,11]
#> 11: (65,4,9), [10,7,18]
#> 12: (89,100,40), [19,17,13]
#> 13: (100,89,39), [8,14,12]
#> 14: (24,100,39), [13,6,23]
#> 15: (41,46,22), [2,27,26]
#> 16: (65,40,64), [19,22,17]
#> 17: (89,40,65), [16,7,12]
#> 18: (65,18,4), [28,11,22]
#> 19: (64,40,100), [12,29,16]
#> 20: (28,7,60), [36,34,25]
#> 21: (46,7,24), [25,6,26]
#> 22: (65,64,18), [41,18,16]
#> 23: (24,28,100), [32,14,25]
#> 24: (14,7,41), [26,47,35]
#> 25: (28,24,7), [21,20,23]
#> 26: (41,7,46), [21,15,24]
#> 27: (41,22,4), [9,28,15]
#> 28: (41,4,18), [18,0,27]
#> 29: (100,52,64), [41,19,30]
#> 30: (100,13,52), [42,29,31]
#> 31: (100,80,13), [37,30,32]
#> 32: (100,28,80), [33,31,23]
#> 33: (85,80,28), [32,34,38]
#> 34: (85,28,60), [20,40,33]
#> 35: (50,7,14), [24,46,36]
#> 36: (50,60,7), [20,35,40]
#> 37: (15,13,80), [31,38,42]
#> 38: (15,80,85), [33,56,37]
#> 39: (58,50,43), [44,48,45]
#> 40: (50,85,60), [34,36,45]
#> 41: (64,52,18), [54,22,29]
#> 42: (52,13,15), [37,53,30]
#> 43: (58,45,53), [57,50,52]
#> 44: (50,26,43), [49,39,46]
#> 45: (58,85,50), [40,39,50]
#> 46: (50,14,26), [47,44,35]
#> 47: (26,14,41), [24,0,46]
#> 48: (3,58,43), [39,49,52]
#> 49: (3,43,26), [44,60,48]
#> 50: (58,53,85), [51,45,43]
#> 51: (85,53,29), [65,56,50]
#> 52: (45,58,3), [48,58,43]
#> 53: (15,27,52), [55,42,61]
#> 54: (84,18,52), [41,55,0]
#> 55: (84,52,27), [53,71,54]
#> 56: (85,29,15), [61,38,51]
#> 57: (8,53,45), [43,58,63]
#> 58: (8,45,3), [52,64,57]
#> 59: (42,3,92), [60,74,64]
#> 60: (3,26,92), [80,59,49]
#> 61: (29,27,15), [53,56,67]
#> 62: (83,8,42), [64,70,63]
#> 63: (83,53,8), [57,62,69]
#> 64: (42,8,3), [58,59,62]
#> 65: (32,29,53), [51,86,85]
#> 66: (27,75,36), [76,71,67]
#> 67: (75,27,29), [61,68,66]
#> 68: (96,75,29), [67,93,78]
#> 69: (73,53,83), [63,87,86]
#> 70: (42,25,83), [91,62,74]
#> 71: (27,36,84), [81,55,66]
#> 72: (86,75,87), [78,79,77]
#> 73: (51,32,73), [86,90,101]
#> 74: (42,92,25), [75,70,59]
#> 75: (92,30,25), [94,74,80]
#> 76: (75,61,36), [82,66,77]
#> 77: (86,61,75), [76,72,83]
#> 78: (96,87,75), [72,68,88]
#> 79: (87,57,86), [100,72,88]
#> 80: (92,26,30), [95,75,60]
#> 81: (37,84,36), [71,82,128]
#> 82: (37,36,61), [76,99,81]
#> 83: (34,61,86), [77,84,99]
#> 84: (86,94,34), [115,83,100]
#> 85: (32,19,29), [96,65,101]
#> 86: (73,32,53), [65,69,73]
#> 87: (83,56,73), [103,69,98]
#> 88: (87,96,57), [89,79,78]
#> 89: (97,57,96), [88,119,116]
#> 90: (73,2,51), [127,73,103]
#> 91: (74,83,25), [70,106,111]
#> 92: (30,54,79), [113,94,109]
#> 93: (29,49,96), [119,68,96]
#> 94: (79,25,30), [75,92,118]
#> 95: (68,30,26), [80,0,109]
#> 96: (29,19,49), [117,93,85]
#> 97: (56,17,90), [135,105,120]
#> 98: (83,81,56), [121,87,111]
#> 99: (34,37,61), [82,83,129]
#> 100: (86,57,94), [116,84,79]
#> 101: (32,51,19), [102,85,73]
#> 102: (82,19,51), [101,108,134]
#> 103: (73,56,2), [105,90,87]
#> 104: (72,69,2), [127,136,144]
#> 105: (90,2,56), [103,97,136]
#> 106: (25,48,74), [133,91,118]
#> 107: (82,69,23), [144,134,108]
#> 108: (51,69,82), [107,102,127]
#> 109: (30,68,54), [125,92,95]
#> 110: (59,81,74), [111,132,130]
#> 111: (74,81,83), [98,91,110]
#> 112: (10,79,33), [113,123,126]
#> 113: (79,54,33), [124,112,92]
#> 114: (21,48,79), [118,126,141]
#> 115: (98,34,94), [84,150,138]
#> 116: (97,94,57), [100,89,147]
#> 117: (78,49,19), [96,143,148]
#> 118: (25,79,48), [114,106,94]
#> 119: (97,96,49), [93,148,89]
#> 120: (56,77,17), [137,97,121]
#> 121: (56,81,77), [131,120,98]
#> 122: (10,55,21), [155,126,123]
#> 123: (55,10,33), [112,124,122]
#> 124: (55,33,54), [113,125,123]
#> 125: (55,54,68), [109,0,124]
#> 126: (21,79,10), [112,122,114]
#> 127: (51,2,69), [104,108,90]
#> 128: (31,84,37), [81,129,0]
#> 129: (34,31,37), [128,99,138]
#> 130: (6,81,59), [110,132,131]
#> 131: (6,77,81), [121,130,140]
#> 132: (6,59,74), [110,133,130]
#> 133: (6,74,48), [106,141,132]
#> 134: (82,23,19), [157,102,107]
#> 135: (90,17,72), [146,136,97]
#> 136: (72,2,90), [105,135,104]
#> 137: (17,77,38), [140,146,120]
#> 138: (98,31,34), [129,115,0]
#> 139: (70,97,78), [148,159,147]
#> 140: (38,77,6), [131,154,137]
#> 141: (48,21,6), [142,133,114]
#> 142: (21,35,6), [154,141,155]
#> 143: (19,91,78), [160,117,157]
#> 144: (23,69,72), [104,152,107]
#> 145: (72,38,93), [153,152,146]
#> 146: (17,38,72), [145,135,137]
#> 147: (70,94,97), [116,139,149]
#> 148: (78,97,49), [119,117,139]
#> 149: (11,94,70), [147,162,150]
#> 150: (11,98,94), [115,149,161]
#> 151: (23,93,62), [168,156,152]
#> 152: (72,93,23), [151,144,145]
#> 153: (38,5,93), [165,145,158]
#> 154: (38,6,35), [142,158,140]
#> 155: (21,55,35), [167,142,122]
#> 156: (23,62,91), [164,157,151]
#> 157: (19,23,91), [156,143,134]
#> 158: (38,35,5), [167,153,154]
#> 159: (63,70,78), [139,160,163]
#> 160: (63,78,91), [143,172,159]
#> 161: (16,98,11), [150,162,0]
#> 162: (70,16,11), [161,149,163]
#> 163: (70,63,16), [180,162,159]
#> 164: (20,91,62), [156,171,172]
#> 165: (12,93,5), [153,176,166]
#> 166: (93,12,99), [175,168,165]
#> 167: (5,35,55), [155,0,158]
#> 168: (93,99,62), [169,151,166]
#> 169: (88,62,99), [168,173,171]
#> 170: (20,88,63), [181,172,171]
#> 171: (88,20,62), [164,169,170]
#> 172: (63,91,20), [164,170,160]
#> 173: (99,66,88), [178,169,174]
#> 174: (99,67,66), [182,173,175]
#> 175: (67,99,12), [166,176,174]
#> 176: (67,12,5), [165,184,175]
#> 177: (47,66,44), [183,185,178]
#> 178: (47,88,66), [173,177,181]
#> 179: (67,76,95), [0,182,184]
#> 180: (47,16,63), [163,181,0]
#> 181: (47,63,88), [170,178,180]
#> 182: (67,95,66), [183,174,179]
#> 183: (95,44,66), [177,182,185]
#> 184: (76,67,5), [176,0,179]
#> 185: (95,47,44), [177,183,0]
#> boundary nodes:  18 41 26 68 55 5 76 95 47 16 98 31 84 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.056477301 0.028372119 0.058880197 0.009375038 0.039677286 0.069126871 
#>                                                                         
#> 0.102352210 0.082333040 0.085279709 0.137730241 0.121563408 0.079639323 
#>                                                                         
#> 0.016684759 0.016908314 0.093831605 0.023856443 0.049565711 0.084525830 
#>                                                                         
#> 0.141961792 0.239849882 0.150618731 0.136777066 0.326273306 0.166369950 
#>                                                                         
#> 0.042853250 0.098709214 0.050209662 0.243100360 0.111999641 0.055382040 
#>                                                                         
#> 0.083154732 0.400648609 0.076897391 0.059392694 0.251247555 0.192201456 
#>                                                                         
#> 0.146132170 0.076060157 0.051514730 0.255235405 0.144668888 0.086882619 
#>                                                                         
#> 0.211190237 0.322818939 0.100729134 0.137822606 0.219492956 0.171618937 
#>                                                                         
#> 0.038444627 0.096165662 0.103662219 0.142475794 0.300286155 0.273957057 
#>                                                                         
#> 0.132445975 0.166743835 0.150616831 0.021387910 0.126036538 0.069033490 
#>                                                                         
#> 0.164166780 0.019848110 0.133531385 0.043590624 0.229950838 0.130731289 
#>                                                                         
#> 0.059883354 0.144298899 0.114391110 0.062962582 0.064364308 0.102771193 
#>                                                                         
#> 0.049493921 0.081147940 0.187983961 0.059473320 0.054327617 0.040363015 
#>                                                                         
#> 0.113483466 0.064252904 0.072375674 0.113241098 0.055094487 0.054892742 
#>                                                                         
#> 0.035836100 0.099923612 0.167540945 0.109750837 0.042174029 0.012162612 
#>                                                                         
#> 0.094908379 0.028798529 0.053486241 0.086580067 0.043662084 0.019692714 
#>                                                                         
#> 0.175737271 0.110180431 0.152346664 0.026103128 0.113638258 0.047446610 
#>                                                                         
#> 0.015029002 0.073970680 0.090970284 0.142990830 0.183154231 0.149426004 
#>                                                                         
#> 0.121224016 0.153212559 0.141034232 0.068808597 0.088362059 0.159157152 
#>                                                                         
#> 0.099769601 0.153281154 0.120633099 0.207379802 0.184882062 0.175086843 
#>                                                                         
#> 0.344844839 0.138087546 0.179684089 0.094886489 0.170789122 0.135262553 
#>                                                                         
#> 0.104153435 0.424923775 0.166992253 0.077493507 0.197976639 0.112202667 
#>                                                                         
#> 0.044423890 0.142512917 0.148094323 0.154580916 0.065241728 0.078455986 
#>                                                                         
#> 0.109025222 0.197607604 0.023725230 0.030943810 0.033221970 0.084271548 
#>                                                                         
#> 0.085038236 0.050849678 0.085831161 0.052109352 0.169049538 0.028404333 
#>                                                                         
#> 0.071643107 0.152540680 0.116283926 0.095881213 0.097186869 0.063402934 
#>                                                                         
#> 0.187805262 0.293266166 0.386616218 0.251945482 0.119786064 0.140730675 
#>                                                                         
#> 0.054902756 0.192905566 0.184388033 0.047961323 0.123376957 0.172393516 
#>                                                                         
#> 0.186386306 0.120101685 0.079681472 0.248169263 0.210183738 0.059575792 
#>                                                                         
#> 0.248504179 0.201190489 0.287627806 0.112963637 0.169004252 0.059331541 
#>                                                                         
#> 0.174304926 0.029390543 0.064201221 0.130679504 0.141861801 0.226948877 
#>                                                                         
#> 0.122973580 0.026849375 0.102442898 0.156987243 0.038876802 0.089136640 
#>                                                                         
#> 0.150191999 0.044706509 0.118514188 0.049186165 0.148535839 0.056715114 
#>                                                                         
#> 0.103751328 0.135997146 0.116594394 0.255134437 0.191951184 0.066527412 
#>                                                                         
#> 0.101525176 0.095568263 0.124991881 0.138261342 0.046002319 0.061579798 
#>                                                                         
#> 0.074957451 0.084951279 0.100265343 0.201508987 0.046941272 0.103193945 
#>                                                                         
#> 0.211870153 0.131108804 0.075705493 0.073725177 0.033736274 0.013547784 
#>                                                                         
#> 0.015675036 0.110452295 0.071800407 0.097155677 0.040596431 0.170138673 
#>                                                                         
#> 0.060523444 0.098389106 0.043685131 0.063704940 0.078618648 0.104733500 
#>                                                                         
#> 0.177457482 0.155224241 0.071187658 0.099485756 0.152688110 0.132177798 
#>                                                                         
#> 0.102837041 0.097231866 0.133103513 0.262479200 0.226758328 0.158750028 
#>                                                                         
#> 0.162737740 0.117312723 0.206485218 0.209205625 0.436944498 0.091423934 
#>                                                                         
#> 0.141595136 0.092882675 0.101666127 0.082177349 0.085593479 0.140181227 
#>                                                                         
#> 0.323566471 0.174228313 0.275009789 0.099174798 0.047179574 0.089274039 
#>                                                                         
#> 0.084845441 0.079032700 0.513309997 0.115547508 0.107201230 0.070662896 
#>                                                                         
#> 0.140309062 0.053552033 0.056976826 0.054486455 0.280392086 0.033441142 
#>                                                                         
#> 0.268729107 0.183546452 0.121755142 0.068328752 0.174979750 0.104632208 
#>                         
#> 0.748401123 0.117340279 
#> 
#> $alpha
#> [1] 0.2970652
#> 
#> attr(,"class")
#> [1] "scree"