Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2         mx1         my1
#>    97   89 0.01158040 0.34049048 0.05012976 0.20828447  0.05080373  0.28020421
#>    89   24 0.05012976 0.20828447 0.08008110 0.18399115  0.08651649  0.22253561
#>    24   12 0.08008110 0.18399115 0.11808146 0.18197979  0.10076246  0.21474764
#>    12   15 0.11808146 0.18197979 0.25530327 0.24252237  0.18416224  0.21798568
#>    15   72 0.25530327 0.24252237 0.26450737 0.26988091  0.18696186  0.28074168
#>    72   59 0.26450737 0.26988091 0.21817377 0.35415930  0.18678005  0.28202445
#>    59   26 0.21817377 0.35415930 0.11696131 0.32064823  0.18548600  0.28328516
#>    26   97 0.11696131 0.32064823 0.01158040 0.34049048  0.05586763  0.28594035
#>    97   48 0.01158040 0.34049048 0.01563722 0.47624928 -2.35955181  0.47928607
#>    48   30 0.01563722 0.47624928 0.07316262 0.52576207 -0.04988332  0.61054670
#>    30    9 0.07316262 0.52576207 0.09993590 0.65693739 -0.03351277  0.61585475
#>     9   10 0.09993590 0.65693739 0.01277186 0.80395800  0.09833271  0.75533566
#>    70   52 0.58261978 0.40204030 0.59484220 0.38441551  0.62595783  0.41904393
#>    52   58 0.59484220 0.38441551 0.68661948 0.36232864  0.64880689  0.40693041
#>    58   66 0.68661948 0.36232864 0.76076275 0.52909979  0.72631016  0.44454984
#>    66    4 0.76076275 0.52909979 0.62752783 0.48930559  0.69647663  0.50139713
#>     4   70 0.62752783 0.48930559 0.58261978 0.40204030  0.58212353  0.45748350
#>    66   56 0.76076275 0.52909979 0.60552073 0.68077617  0.66966973  0.59114927
#>    56   42 0.60552073 0.68077617 0.56799927 0.54610068  0.66761144  0.59091269
#>    42   77 0.56799927 0.54610068 0.51350895 0.52653202  0.56120091  0.47938079
#>    77   79 0.51350895 0.52653202 0.54663249 0.40884686  0.56014767  0.47615487
#>    79   70 0.54663249 0.40884686 0.58261978 0.40204030  0.57514784  0.46107330
#>    74   92 0.88827318 1.00000000 0.78315185 0.94078571  0.80871181  1.01832639
#>    92   73 0.78315185 0.94078571 0.84710004 0.90553939  0.82061256  0.93311704
#>    73   44 0.84710004 0.90553939 1.00000000 0.90311232  0.92373774  0.91615174
#>    44   74 1.00000000 0.90311232 0.88827318 1.00000000  0.92999121  0.93524431
#>    78   11 0.55931290 0.02403652 0.51789052 0.02943382  0.44595947 -0.68426133
#>    11   55 0.51789052 0.02943382 0.38951596 0.06685309  0.39243277 -0.16205762
#>    55   82 0.38951596 0.06685309 0.17707763 0.04287172  0.27677480  0.11263732
#>    82   88 0.17707763 0.04287172 0.06938922 0.06117343  0.09483452 -0.11507835
#>    52   86 0.59484220 0.38441551 0.55883075 0.32634395  0.60579418  0.33742242
#>    86   28 0.55883075 0.32634395 0.58446970 0.21304931  0.56971850  0.26925947
#>    28   21 0.58446970 0.21304931 0.69285214 0.29490688  0.66044859  0.22513042
#>    21   58 0.69285214 0.29490688 0.68661948 0.36232864  0.65678221  0.32557144
#>    76   90 0.81000278 0.65499210 0.87130884 0.53517499  0.87805209  0.61421786
#>    90   38 0.87130884 0.53517499 0.96617058 0.33719158  1.00606373  0.47802370
#>    38   19 0.96617058 0.33719158 0.99325594 0.64925890  1.11515265  0.48147000
#>    19   37 0.99325594 0.64925890 0.93481422 0.69008732  0.95415355  0.65552870
#>    37   76 0.93481422 0.69008732 0.81000278 0.65499210  0.88326267  0.63393833
#>    30   18 0.07316262 0.52576207 0.24187591 0.54670249  0.15234497  0.57792062
#>    18   25 0.24187591 0.54670249 0.25440666 0.71298060  0.19573548  0.63379086
#>    25   10 0.25440666 0.71298060 0.01277186 0.80395800  0.14220190  0.78134437
#>    25   49 0.25440666 0.71298060 0.42170633 0.80633430  0.21623480  0.97797478
#>    49   34 0.42170633 0.80633430 0.48229968 0.92258788  0.21618017  0.98737598
#>    34   17 0.48229968 0.92258788 0.58211118 0.97561563 -0.05195681  2.04864125
#>    78   39 0.55931290 0.02403652 0.51947398 0.08654457  0.54258772  0.05732639
#>    39   55 0.51947398 0.08654457 0.38951596 0.06685309  0.45707105  0.05969743
#>    24   40 0.08008110 0.18399115 0.07993559 0.08539462  0.09652385  0.13466851
#>    40   32 0.07993559 0.08539462 0.16543167 0.05309614  0.14079385  0.11718428
#>    32   15 0.16543167 0.05309614 0.25530327 0.24252237  0.21638357  0.14495497
#>    92   84 0.78315185 0.94078571 0.64090254 0.89619192  0.70708776  0.93424507
#>    84   87 0.64090254 0.89619192 0.73471787 0.80070554  0.67771401  0.83852923
#>    87   27 0.73471787 0.80070554 0.76469743 0.85572093  0.74121871  0.83283912
#>    27   92 0.76469743 0.85572093 0.78315185 0.94078571  0.76849043  0.89943225
#>    88   40 0.06938922 0.06117343 0.07993559 0.08539462  0.05236049  0.08299470
#>    32   82 0.16543167 0.05309614 0.17707763 0.04287172  0.24462553  0.13155582
#>    55   41 0.38951596 0.06685309 0.43326658 0.16678382  0.32331468  0.15537921
#>    41   99 0.43326658 0.16678382 0.42595367 0.27828823  0.34966421  0.21729285
#>    99   72 0.42595367 0.27828823 0.26450737 0.26988091  0.34769988  0.22666532
#>    84   65 0.64090254 0.89619192 0.48052363 0.88685914  0.56290367  0.85388141
#>    65   63 0.48052363 0.88685914 0.46318040 0.84608837  0.51885694  0.84647861
#>    63   49 0.46318040 0.84608837 0.42170633 0.80633430  0.45977445  0.80813040
#>    49   69 0.42170633 0.80633430 0.42209811 0.67295935  0.45004006  0.73972948
#>    69   56 0.42209811 0.67295935 0.60552073 0.68077617  0.51273435  0.70209437
#>    56   31 0.60552073 0.68077617 0.63221084 0.74485160  0.55281891  0.74032518
#>    31   84 0.63221084 0.74485160 0.64090254 0.89619192  0.58389922  0.82354596
#>    34   65 0.48229968 0.92258788 0.48052363 0.88685914  0.55780005  0.90092628
#>    30   50 0.07316262 0.52576207 0.09275120 0.44879157  0.16112636  0.50717052
#>    50   20 0.09275120 0.44879157 0.17027207 0.39438917  0.17584152  0.48475838
#>    20   59 0.17027207 0.39438917 0.21817377 0.35415930  0.25554419  0.44728945
#>    59   33 0.21817377 0.35415930 0.38780006 0.50540891  0.29130307  0.44288752
#>    33   64 0.38780006 0.50540891 0.33823762 0.61004647  0.31773023  0.53627637
#>    64   85 0.33823762 0.61004647 0.32850499 0.64411696  0.26994841  0.60896423
#>    85   25 0.32850499 0.64411696 0.25440666 0.71298060  0.24631727  0.62997900
#>    96   44 0.99721005 0.14076486 1.00000000 0.90311232 12.09538935  0.48132782
#>    44   35 1.00000000 0.90311232 0.92412569 0.77004814  1.07146642  0.77419740
#>    35   37 0.92412569 0.77004814 0.93481422 0.69008732  1.01413428  0.74138499
#>    38   96 0.96617058 0.33719158 0.99721005 0.14076486  2.01952099  0.40297685
#>    50   26 0.09275120 0.44879157 0.11696131 0.32064823  0.10575607  0.38488990
#>    28   98 0.58446970 0.21304931 0.76735153 0.20976205  0.67579215  0.20481498
#>    98   93 0.76735153 0.20976205 0.76607184 0.25089448  0.70267412  0.22833597
#>    93   80 0.76607184 0.25089448 0.82673930 0.38088506  0.76430747  0.33087016
#>    80   58 0.82673930 0.38088506 0.68661948 0.36232864  0.76149405  0.33525129
#>    35   67 0.92412569 0.77004814 0.90399998 0.79668196  0.85307561  0.73728036
#>    67   87 0.90399998 0.79668196 0.73471787 0.80070554  0.81848225  0.76181010
#>    87   43 0.73471787 0.80070554 0.68974201 0.73231408  0.76680571  0.73061948
#>    43   76 0.68974201 0.73231408 0.81000278 0.65499210  0.74167593  0.68090494
#>    98   91 0.76735153 0.20976205 0.84789420 0.11589828  0.80423496  0.15992307
#>    91   47 0.84789420 0.11589828 0.85374017 0.11574659  0.85209288  0.16498606
#>    47   53 0.85374017 0.11574659 0.99282890 0.15250663  0.91247486  0.17502713
#>    53   46 0.99282890 0.15250663 0.96811211 0.26445626  0.92021805  0.19517862
#>    46   60 0.96811211 0.26445626 0.84569162 0.25263175  0.91074245  0.21878205
#>    60   93 0.84569162 0.25263175 0.76607184 0.25089448  0.80615220  0.23936746
#>    85   54 0.32850499 0.64411696 0.35416351 0.74306667  0.31240725  0.70109284
#>    54   49 0.35416351 0.74306667 0.42170633 0.80633430  0.24379996  0.92857510
#>    50   97 0.09275120 0.44879157 0.01158040 0.34049048  0.07334596  0.37876666
#>    67   73 0.90399998 0.79668196 0.84710004 0.90553939  0.83618519  0.83053462
#>    86   51 0.55883075 0.32634395 0.56357792 0.21024336  0.56638712  0.26850557
#>    51   95 0.56357792 0.21024336 0.55926828 0.17471732  0.57684930  0.19060900
#>    95   94 0.55926828 0.17471732 0.56564811 0.12851876  0.62336270  0.16002870
#>    94   98 0.56564811 0.12851876 0.76735153 0.20976205  0.67478133  0.14857982
#>    99   57 0.42595367 0.27828823 0.47221085 0.29412174  0.44131836  0.30888701
#>    57   79 0.47221085 0.29412174 0.54663249 0.40884686  0.47501228  0.37380552
#>    79   68 0.54663249 0.40884686 0.44716445 0.47475611  0.46385483  0.39193313
#>    68   33 0.44716445 0.47475611 0.38780006 0.50540891  0.37748145  0.41261408
#>    48   50 0.01563722 0.47624928 0.09275120 0.44879157  0.06102141  0.48169435
#>    56   43 0.60552073 0.68077617 0.68974201 0.73231408  0.65232356  0.69887731
#>    68   77 0.44716445 0.47475611 0.51350895 0.52653202  0.44579190  0.54490900
#>    77   69 0.51350895 0.52653202 0.42209811 0.67295935  0.43954215  0.58210283
#>    69   64 0.42209811 0.67295935 0.33823762 0.61004647  0.42204299  0.58568496
#>    44   67 1.00000000 0.90311232 0.90399998 0.79668196  0.97484270  0.82929305
#>    83   96 0.99792897 0.05982255 0.99721005 0.14076486  1.10742875  0.10126946
#>    96   61 0.99721005 0.14076486 0.92204966 0.07105789  0.94502764  0.12165597
#>    61   83 0.92204966 0.07105789 0.99792897 0.05982255  0.96023541  0.06710227
#>   100   46 0.87595923 0.31526447 0.96811211 0.26445626  0.92749694  0.29976569
#>    46   96 0.96811211 0.26445626 0.99721005 0.14076486  1.45557497  0.31386176
#>    90   81 0.87130884 0.53517499 0.86587913 0.36008062  0.93812517  0.44547163
#>    81  100 0.86587913 0.36008062 0.87595923 0.31526447  0.91582798  0.34777349
#>    78   94 0.55931290 0.02403652 0.56564811 0.12851876  0.57146760  0.07573271
#>    94   75 0.56564811 0.12851876 0.49000661 0.14918367  0.52465295  0.12723164
#>    75   45 0.49000661 0.14918367 0.44880701 0.13352571  0.48233943  0.10732606
#>    45   55 0.44880701 0.13352571 0.38951596 0.06685309  0.45588384  0.06753272
#>    75   57 0.49000661 0.14918367 0.47221085 0.29412174  0.48139700  0.22168810
#>    56   71 0.60552073 0.68077617 0.55230052 0.54959501  0.57545078  0.61658924
#>    71   77 0.55230052 0.54959501 0.51350895 0.52653202  0.55114364  0.50738596
#>    66   76 0.76076275 0.52909979 0.81000278 0.65499210  0.70480680  0.62356147
#>   100   60 0.87595923 0.31526447 0.84569162 0.25263175  0.90658399  0.26183503
#>    53   96 0.99282890 0.15250663 0.99721005 0.14076486  1.42207677  0.30598143
#>    91   61 0.84789420 0.11589828 0.92204966 0.07105789  0.84868005  0.03345984
#>    61   53 0.92204966 0.07105789 0.99282890 0.15250663  0.92076755  0.14365013
#>    54   69 0.35416351 0.74306667 0.42209811 0.67295935  0.42077293  0.73964351
#>    86   57 0.55883075 0.32634395 0.47221085 0.29412174  0.53157269  0.26708207
#>    75   95 0.49000661 0.14918367 0.55926828 0.17471732  0.51113345  0.19858095
#>    79   86 0.54663249 0.40884686 0.55883075 0.32634395  0.55628938  0.36812143
#>    85   69 0.32850499 0.64411696 0.42209811 0.67295935  0.36651103  0.68706326
#>    69   71 0.42209811 0.67295935 0.55230052 0.54959501  0.51533897  0.64097661
#>    80   66 0.82673930 0.38088506 0.76076275 0.52909979  0.74831494  0.43476694
#>   100   93 0.87595923 0.31526447 0.76607184 0.25089448  0.80458646  0.31112590
#>    91   62 0.84789420 0.11589828 0.62579750 0.03522434  0.78836848 -0.06628133
#>    62   83 0.62579750 0.03522434 0.99792897 0.05982255  0.85676179 -0.63171977
#>    78   83 0.55931290 0.02403652 0.99792897 0.05982255  0.95568833 -2.12831987
#>    80   90 0.82673930 0.38088506 0.87130884 0.53517499  0.81964201  0.46651758
#>     1   89 0.11548870 0.24875967 0.05012976 0.20828447  0.05080373  0.28020421
#>     1   24 0.11548870 0.24875967 0.08008110 0.18399115  0.08651649  0.22253561
#>     1   12 0.11548870 0.24875967 0.11808146 0.18197979  0.10076246  0.21474764
#>     1   15 0.11548870 0.24875967 0.25530327 0.24252237  0.18416224  0.21798568
#>     1   72 0.11548870 0.24875967 0.26450737 0.26988091  0.18696186  0.28074168
#>     1   59 0.11548870 0.24875967 0.21817377 0.35415930  0.18678005  0.28202445
#>     1   26 0.11548870 0.24875967 0.11696131 0.32064823  0.18548600  0.28328516
#>     1   97 0.11548870 0.24875967 0.01158040 0.34049048  0.05586763  0.28594035
#>     2   48 0.00000000 0.75140262 0.01563722 0.47624928 -2.35955181  0.47928607
#>     2   30 0.00000000 0.75140262 0.07316262 0.52576207 -0.04988332  0.61054670
#>     2    9 0.00000000 0.75140262 0.09993590 0.65693739 -0.03351277  0.61585475
#>     2   10 0.00000000 0.75140262 0.01277186 0.80395800  0.09833271  0.75533566
#>     3   52 0.63715682 0.46423124 0.59484220 0.38441551  0.62595783  0.41904393
#>     3   58 0.63715682 0.46423124 0.68661948 0.36232864  0.64880689  0.40693041
#>     3   66 0.63715682 0.46423124 0.76076275 0.52909979  0.72631016  0.44454984
#>     3    4 0.63715682 0.46423124 0.62752783 0.48930559  0.69647663  0.50139713
#>     3   70 0.63715682 0.46423124 0.58261978 0.40204030  0.58212353  0.45748350
#>     4   56 0.62752783 0.48930559 0.60552073 0.68077617  0.66966973  0.59114927
#>     4   42 0.62752783 0.48930559 0.56799927 0.54610068  0.66761144  0.59091269
#>     4   77 0.62752783 0.48930559 0.51350895 0.52653202  0.56120091  0.47938079
#>     4   79 0.62752783 0.48930559 0.54663249 0.40884686  0.56014767  0.47615487
#>     5   92 0.85453869 0.95075594 0.78315185 0.94078571  0.80871181  1.01832639
#>     5   73 0.85453869 0.95075594 0.84710004 0.90553939  0.82061256  0.93311704
#>     5   44 0.85453869 0.95075594 1.00000000 0.90311232  0.92373774  0.91615174
#>     5   74 0.85453869 0.95075594 0.88827318 1.00000000  0.92999121  0.93524431
#>     6   11 0.23073568 0.00000000 0.51789052 0.02943382  0.44595947 -0.68426133
#>     6   55 0.23073568 0.00000000 0.38951596 0.06685309  0.39243277 -0.16205762
#>     6   82 0.23073568 0.00000000 0.17707763 0.04287172  0.27677480  0.11263732
#>     6   88 0.23073568 0.00000000 0.06938922 0.06117343  0.09483452 -0.11507835
#>     7   52 0.62304398 0.29235869 0.59484220 0.38441551  0.63417143  0.34611585
#>     7   86 0.62304398 0.29235869 0.55883075 0.32634395  0.60579418  0.33742242
#>     7   28 0.62304398 0.29235869 0.58446970 0.21304931  0.56971850  0.26925947
#>     7   21 0.62304398 0.29235869 0.69285214 0.29490688  0.66044859  0.22513042
#>     7   58 0.62304398 0.29235869 0.68661948 0.36232864  0.65678221  0.32557144
#>     8   90 0.95736076 0.61605692 0.87130884 0.53517499  0.87805209  0.61421786
#>     8   38 0.95736076 0.61605692 0.96617058 0.33719158  1.00606373  0.47802370
#>     8   19 0.95736076 0.61605692 0.99325594 0.64925890  1.11515265  0.48147000
#>     8   37 0.95736076 0.61605692 0.93481422 0.69008732  0.95415355  0.65552870
#>     8   76 0.95736076 0.61605692 0.81000278 0.65499210  0.88326267  0.63393833
#>     9   18 0.09993590 0.65693739 0.24187591 0.54670249  0.15234497  0.57792062
#>     9   25 0.09993590 0.65693739 0.25440666 0.71298060  0.19573548  0.63379086
#>    10   49 0.01277186 0.80395800 0.42170633 0.80633430  0.21623480  0.97797478
#>    10   34 0.01277186 0.80395800 0.48229968 0.92258788  0.21618017  0.98737598
#>    10   17 0.01277186 0.80395800 0.58211118 0.97561563 -0.05195681  2.04864125
#>    11   39 0.51789052 0.02943382 0.51947398 0.08654457  0.54258772  0.05732639
#>    12   40 0.11808146 0.18197979 0.07993559 0.08539462  0.09652385  0.13466851
#>    12   32 0.11808146 0.18197979 0.16543167 0.05309614  0.14079385  0.11718428
#>    13   84 0.74053398 0.86561653 0.64090254 0.89619192  0.70708776  0.93424507
#>    13   87 0.74053398 0.86561653 0.73471787 0.80070554  0.67771401  0.83852923
#>    13   27 0.74053398 0.86561653 0.76469743 0.85572093  0.74121871  0.83283912
#>    13   92 0.74053398 0.86561653 0.78315185 0.94078571  0.76849043  0.89943225
#>    14   40 0.04781754 0.11029868 0.07993559 0.08539462  0.05236049  0.08299470
#>    14   24 0.04781754 0.11029868 0.08008110 0.18399115  0.09243260  0.13467455
#>    14   89 0.04781754 0.11029868 0.05012976 0.20828447  0.03547801  0.15961004
#>    14   97 0.04781754 0.11029868 0.01158040 0.34049048 -0.33296007  0.16830424
#>    15   82 0.25530327 0.24252237 0.17707763 0.04287172  0.24462553  0.13155582
#>    15   55 0.25530327 0.24252237 0.38951596 0.06685309  0.27602811  0.11925190
#>    15   41 0.25530327 0.24252237 0.43326658 0.16678382  0.32331468  0.15537921
#>    15   99 0.25530327 0.24252237 0.42595367 0.27828823  0.34966421  0.21729285
#>    16   65 0.49583010 0.79578547 0.48052363 0.88685914  0.56290367  0.85388141
#>    16   63 0.49583010 0.79578547 0.46318040 0.84608837  0.51885694  0.84647861
#>    16   49 0.49583010 0.79578547 0.42170633 0.80633430  0.45977445  0.80813040
#>    16   69 0.49583010 0.79578547 0.42209811 0.67295935  0.45004006  0.73972948
#>    16   56 0.49583010 0.79578547 0.60552073 0.68077617  0.51273435  0.70209437
#>    16   31 0.49583010 0.79578547 0.63221084 0.74485160  0.55281891  0.74032518
#>    16   84 0.49583010 0.79578547 0.64090254 0.89619192  0.58389922  0.82354596
#>    17   65 0.58211118 0.97561563 0.48052363 0.88685914  0.55780005  0.90092628
#>    17   84 0.58211118 0.97561563 0.64090254 0.89619192  0.56033486  0.89802502
#>    17   92 0.58211118 0.97561563 0.78315185 0.94078571  0.68866336  0.99301689
#>    17   74 0.58211118 0.97561563 0.88827318 1.00000000  0.72025435  1.17536224
#>    18   50 0.24187591 0.54670249 0.09275120 0.44879157  0.16112636  0.50717052
#>    18   20 0.24187591 0.54670249 0.17027207 0.39438917  0.17584152  0.48475838
#>    18   59 0.24187591 0.54670249 0.21817377 0.35415930  0.25554419  0.44728945
#>    18   33 0.24187591 0.54670249 0.38780006 0.50540891  0.29130307  0.44288752
#>    18   64 0.24187591 0.54670249 0.33823762 0.61004647  0.31773023  0.53627637
#>    18   85 0.24187591 0.54670249 0.32850499 0.64411696  0.26994841  0.60896423
#>    19   44 0.99325594 0.64925890 1.00000000 0.90311232 12.09538935  0.48132782
#>    19   35 0.99325594 0.64925890 0.92412569 0.77004814  1.07146642  0.77419740
#>    19   96 0.99325594 0.64925890 0.99721005 0.14076486  2.01952099  0.40297685
#>    20   26 0.17027207 0.39438917 0.11696131 0.32064823  0.10575607  0.38488990
#>    21   98 0.69285214 0.29490688 0.76735153 0.20976205  0.67579215  0.20481498
#>    21   93 0.69285214 0.29490688 0.76607184 0.25089448  0.70267412  0.22833597
#>    21   80 0.69285214 0.29490688 0.82673930 0.38088506  0.76430747  0.33087016
#>    22   35 0.81386719 0.66957112 0.92412569 0.77004814  0.86723004  0.72174798
#>    22   67 0.81386719 0.66957112 0.90399998 0.79668196  0.85307561  0.73728036
#>    22   87 0.81386719 0.66957112 0.73471787 0.80070554  0.81848225  0.76181010
#>    22   43 0.81386719 0.66957112 0.68974201 0.73231408  0.76680571  0.73061948
#>    22   76 0.81386719 0.66957112 0.81000278 0.65499210  0.74167593  0.68090494
#>    22   37 0.81386719 0.66957112 0.93481422 0.69008732  0.88039550  0.64413502
#>    23   91 0.83771589 0.21210868 0.84789420 0.11589828  0.80423496  0.15992307
#>    23   47 0.83771589 0.21210868 0.85374017 0.11574659  0.85209288  0.16498606
#>    23   53 0.83771589 0.21210868 0.99282890 0.15250663  0.91247486  0.17502713
#>    23   46 0.83771589 0.21210868 0.96811211 0.26445626  0.92021805  0.19517862
#>    23   60 0.83771589 0.21210868 0.84569162 0.25263175  0.91074245  0.21878205
#>    23   93 0.83771589 0.21210868 0.76607184 0.25089448  0.80615220  0.23936746
#>    23   98 0.83771589 0.21210868 0.76735153 0.20976205  0.80185050  0.23142148
#>    25   54 0.25440666 0.71298060 0.35416351 0.74306667  0.31240725  0.70109284
#>    27   67 0.76469743 0.85572093 0.90399998 0.79668196  0.81916098  0.79036599
#>    27   73 0.76469743 0.85572093 0.84710004 0.90553939  0.83618519  0.83053462
#>    28   51 0.58446970 0.21304931 0.56357792 0.21024336  0.56638712  0.26850557
#>    28   95 0.58446970 0.21304931 0.55926828 0.17471732  0.57684930  0.19060900
#>    28   94 0.58446970 0.21304931 0.56564811 0.12851876  0.62336270  0.16002870
#>    29   57 0.41146320 0.32565091 0.47221085 0.29412174  0.44131836  0.30888701
#>    29   79 0.41146320 0.32565091 0.54663249 0.40884686  0.47501228  0.37380552
#>    29   68 0.41146320 0.32565091 0.44716445 0.47475611  0.46385483  0.39193313
#>    29   33 0.41146320 0.32565091 0.38780006 0.50540891  0.37748145  0.41261408
#>    29   59 0.41146320 0.32565091 0.21817377 0.35415930  0.32451399  0.40564151
#>    29   72 0.41146320 0.32565091 0.26450737 0.26988091  0.31682654  0.35351992
#>    29   99 0.41146320 0.32565091 0.42595367 0.27828823  0.34495349  0.27940447
#>    31   43 0.63221084 0.74485160 0.68974201 0.73231408  0.65232356  0.69887731
#>    31   87 0.63221084 0.74485160 0.73471787 0.80070554  0.67272413  0.79248977
#>    32   88 0.16543167 0.05309614 0.06938922 0.06117343  0.11721053  0.05475775
#>    33   77 0.38780006 0.50540891 0.51350895 0.52653202  0.44579190  0.54490900
#>    33   69 0.38780006 0.50540891 0.42209811 0.67295935  0.43954215  0.58210283
#>    34   63 0.48229968 0.92258788 0.46318040 0.84608837  0.35963040  0.91260726
#>    36   96 0.99086784 0.09024212 0.99721005 0.14076486  1.10742875  0.10126946
#>    36   61 0.99086784 0.09024212 0.92204966 0.07105789  0.94502764  0.12165597
#>    36   83 0.99086784 0.09024212 0.99792897 0.05982255  0.96023541  0.06710227
#>    38   46 0.96617058 0.33719158 0.96811211 0.26445626  0.92749694  0.29976569
#>    38   81 0.96617058 0.33719158 0.86587913 0.36008062  0.93812517  0.44547163
#>    38  100 0.96617058 0.33719158 0.87595923 0.31526447  0.91582798  0.34777349
#>    39   94 0.51947398 0.08654457 0.56564811 0.12851876  0.57146760  0.07573271
#>    39   75 0.51947398 0.08654457 0.49000661 0.14918367  0.52465295  0.12723164
#>    39   45 0.51947398 0.08654457 0.44880701 0.13352571  0.48233943  0.10732606
#>    41   57 0.43326658 0.16678382 0.47221085 0.29412174  0.48139700  0.22168810
#>    41   45 0.43326658 0.16678382 0.44880701 0.13352571  0.38986391  0.12624333
#>    41   75 0.43326658 0.16678382 0.49000661 0.14918367  0.46228836  0.16008493
#>    42   71 0.56799927 0.54610068 0.55230052 0.54959501  0.57545078  0.61658924
#>    43   66 0.68974201 0.73231408 0.76076275 0.52909979  0.69953860  0.62172030
#>    44   83 1.00000000 0.90311232 0.99792897 0.05982255 34.61758296  0.39890361
#>    47   61 0.85374017 0.11574659 0.92204966 0.07105789  0.84868005  0.03345984
#>    51   57 0.56357792 0.21024336 0.47221085 0.29412174  0.53157269  0.26708207
#>    51   75 0.56357792 0.21024336 0.49000661 0.14918367  0.49097946  0.22286465
#>    52   79 0.59484220 0.38441551 0.54663249 0.40884686  0.55832792  0.37214402
#>    61   62 0.92204966 0.07105789 0.62579750 0.03522434  0.78836848 -0.06628133
#>    62   98 0.62579750 0.03522434 0.76735153 0.20976205  0.72949726  0.09579211
#>    62   94 0.62579750 0.03522434 0.56564811 0.12851876  0.67981329  0.13608692
#>    62   78 0.62579750 0.03522434 0.55931290 0.02403652  0.58493466  0.07491615
#>    66   90 0.76076275 0.52909979 0.87130884 0.53517499  0.81964201  0.46651758
#>    80   81 0.82673930 0.38088506 0.86587913 0.36008062  0.82315925  0.32693032
#>    80  100 0.82673930 0.38088506 0.87595923 0.31526447  0.80382398  0.31242755
#>     2   97 0.00000000 0.75140262 0.01158040 0.34049048 -2.35955181  0.47928607
#>     6   78 0.23073568 0.00000000 0.55931290 0.02403652  0.44595947 -0.68426133
#>    14   88 0.04781754 0.11029868 0.06938922 0.06117343  0.05236049  0.08299470
#>           mx2         my2 bp1 bp2
#>   -0.33296007  0.16830424   0   0
#>    0.03547801  0.15961004   0   0
#>    0.09652385  0.13466851   0   0
#>    0.21638357  0.14495497   0   0
#>    0.34769988  0.22666532   0   0
#>    0.31682654  0.35351992   0   0
#>    0.16634883  0.34108457   0   0
#>    0.07334596  0.37876666   0   0
#>    0.03468871  0.40773996   0   0
#>    0.06102141  0.48169435   0   0
#>    0.15234497  0.57792062   0   0
#>    0.14220190  0.78134437   0   0
#>    0.55832792  0.37214402   0   0
#>    0.63417143  0.34611585   0   0
#>    0.74831494  0.43476694   0   0
#>    0.66966973  0.59114927   0   0
#>    0.57514784  0.46107330   0   0
#>    0.69953860  0.62172030   0   0
#>    0.57545078  0.61658924   0   0
#>    0.55114364  0.50738596   0   0
#>    0.51037996  0.46214730   0   0
#>    0.55832792  0.37214402   0   0
#>    0.72025435  1.17536224   0   0
#>    0.79846262  0.89292992   0   0
#>    0.92309983  0.87596504   0   0
#>    5.36013360  6.04389940   0   1
#>    0.54258772  0.05732639   0   0
#>    0.45707105  0.05969743   0   0
#>    0.27602811  0.11925190   0   0
#>    0.11085491 -0.02081340   0   0
#>    0.55628938  0.36812143   0   0
#>    0.56638712  0.26850557   0   0
#>    0.67579215  0.20481498   0   0
#>    0.76149405  0.33525129   0   0
#>    0.81334448  0.58110933   0   0
#>    0.93812517  0.44547163   0   0
#>    2.01952099  0.40297685   0   0
#>    1.01413428  0.74138499   0   0
#>    0.88039550  0.64413502   0   0
#>    0.16112636  0.50717052   0   0
#>    0.24631727  0.62997900   0   0
#>    0.21623480  0.97797478   0   0
#>    0.24379996  0.92857510   0   0
#>    0.35963040  0.91260726   0   0
#>    0.55780005  0.90092628   0   0
#>    0.57146760  0.07573271   0   0
#>    0.45588384  0.06753272   0   0
#>    0.09243260  0.13467455   0   0
#>    0.11721053  0.05475775   0   0
#>    0.24462553  0.13155582   0   0
#>    0.68866336  0.99301689   0   0
#>    0.65812503  0.81928306   0   0
#>    0.81916098  0.79036599   0   0
#>    0.79846262  0.89292992   0   0
#>    0.11721053  0.05475775   0   0
#>    0.11085491 -0.02081340   0   0
#>    0.38986391  0.12624333   0   0
#>    0.46996967  0.22518297   0   0
#>    0.34495349  0.27940447   0   0
#>    0.56033486  0.89802502   0   0
#>    0.36877081  0.91032283   0   0
#>    0.35963040  0.91260726   0   0
#>    0.42077293  0.73964351   0   0
#>    0.51533897  0.64097661   0   0
#>    0.65232356  0.69887731   0   0
#>    0.65812503  0.81928306   0   0
#>    0.36877081  0.91032283   0   0
#>    0.06102141  0.48169435   0   0
#>    0.10575607  0.38488990   0   0
#>    0.16634883  0.34108457   0   0
#>    0.32451399  0.40564151   0   0
#>    0.42204299  0.58568496   0   0
#>    0.38081855  0.64063557   0   0
#>    0.31240725  0.70109284   0   0
#>   34.61758296  0.39890361   0   0
#>    0.97484270  0.82929305   0   0
#>    0.86723004  0.72174798   0   0
#>    1.45557497  0.31386176   0   0
#>    0.07334596  0.37876666   0   0
#>    0.67478133  0.14857982   0   0
#>    0.80185050  0.23142148   0   0
#>    0.80382398  0.31242755   0   0
#>    0.74831494  0.43476694   0   0
#>    0.97484270  0.82929305   0   0
#>    0.81916098  0.79036599   0   0
#>    0.67272413  0.79248977   0   0
#>    0.70480680  0.62356147   0   0
#>    0.72949726  0.09579211   0   0
#>    0.84868005  0.03345984   0   0
#>    0.92076755  0.14365013   0   0
#>    1.42207677  0.30598143   0   0
#>    0.90658399  0.26183503   0   0
#>    0.80458646  0.31112590   0   0
#>    0.36651103  0.68706326   0   0
#>    0.42077293  0.73964351   0   0
#>    0.03468871  0.40773996   0   0
#>    0.92309983  0.87596504   0   0
#>    0.53157269  0.26708207   0   0
#>    0.51113345  0.19858095   0   0
#>    0.53009420  0.14714870   0   0
#>    0.67981329  0.13608692   0   0
#>    0.46996967  0.22518297   0   0
#>    0.49723455  0.35939003   0   0
#>    0.51037996  0.46214730   0   0
#>    0.44579190  0.54490900   0   0
#>    0.03468871  0.40773996   0   0
#>    0.69953860  0.62172030   0   0
#>    0.51037996  0.46214730   0   0
#>    0.48853711  0.61268913   0   0
#>    0.38081855  0.64063557   0   0
#>    0.92309983  0.87596504   0   0
#>   34.61758296  0.39890361   0   0
#>    0.94066857  0.12635607   0   0
#>    0.85676179 -0.63171977   0   0
#>    0.90658399  0.26183503   0   0
#>    1.42207677  0.30598143   0   0
#>    0.88701137  0.44705668   0   0
#>    0.82315925  0.32693032   0   0
#>    0.58493466  0.07491615   0   0
#>    0.53009420  0.14714870   0   0
#>    0.46228836  0.16008493   0   0
#>    0.38986391  0.12624333   0   0
#>    0.49097946  0.22286465   0   0
#>    0.51533897  0.64097661   0   0
#>    0.48853711  0.61268913   0   0
#>    0.81334448  0.58110933   0   0
#>    0.80458646  0.31112590   0   0
#>    0.94066857  0.12635607   0   0
#>    0.78836848 -0.06628133   0   0
#>    0.94066857  0.12635607   0   0
#>    0.36651103  0.68706326   0   0
#>    0.49723455  0.35939003   0   0
#>    0.53009420  0.14714870   0   0
#>    0.49723455  0.35939003   0   0
#>    0.38081855  0.64063557   0   0
#>    0.48853711  0.61268913   0   0
#>    0.81964201  0.46651758   0   0
#>    0.80382398  0.31242755   0   0
#>    0.72949726  0.09579211   0   0
#>    0.95568833 -2.12831987   0   0
#>    1.14047176 -4.39314189   0   1
#>    0.88701137  0.44705668   0   0
#>    0.08651649  0.22253561   0   0
#>    0.10076246  0.21474764   0   0
#>    0.18416224  0.21798568   0   0
#>    0.18696186  0.28074168   0   0
#>    0.18678005  0.28202445   0   0
#>    0.18548600  0.28328516   0   0
#>    0.05586763  0.28594035   0   0
#>    0.05080373  0.28020421   0   0
#>   -0.04988332  0.61054670   0   0
#>   -0.03351277  0.61585475   0   0
#>    0.09833271  0.75533566   0   0
#>  -17.86816450  5.12150318   0   1
#>    0.64880689  0.40693041   0   0
#>    0.72631016  0.44454984   0   0
#>    0.69647663  0.50139713   0   0
#>    0.58212353  0.45748350   0   0
#>    0.62595783  0.41904393   0   0
#>    0.66761144  0.59091269   0   0
#>    0.56120091  0.47938079   0   0
#>    0.56014767  0.47615487   0   0
#>    0.57514784  0.46107330   0   0
#>    0.82061256  0.93311704   0   0
#>    0.92373774  0.91615174   0   0
#>    0.92999121  0.93524431   0   0
#>    0.80871181  1.01832639   0   0
#>    0.39243277 -0.16205762   0   0
#>    0.27677480  0.11263732   0   0
#>    0.09483452 -0.11507835   0   0
#>   -1.95969752 -5.53395893   0   1
#>    0.60579418  0.33742242   0   0
#>    0.56971850  0.26925947   0   0
#>    0.66044859  0.22513042   0   0
#>    0.65678221  0.32557144   0   0
#>    0.63417143  0.34611585   0   0
#>    1.00606373  0.47802370   0   0
#>    1.11515265  0.48147000   0   0
#>    0.95415355  0.65552870   0   0
#>    0.88326267  0.63393833   0   0
#>    0.87805209  0.61421786   0   0
#>    0.19573548  0.63379086   0   0
#>    0.14220190  0.78134437   0   0
#>    0.21618017  0.98737598   0   0
#>   -0.05195681  2.04864125   0   0
#>   -0.53739524  3.65870181   0   1
#>    0.45707105  0.05969743   0   0
#>    0.14079385  0.11718428   0   0
#>    0.21638357  0.14495497   0   0
#>    0.67771401  0.83852923   0   0
#>    0.74121871  0.83283912   0   0
#>    0.76849043  0.89943225   0   0
#>    0.70708776  0.93424507   0   0
#>    0.09243260  0.13467455   0   0
#>    0.03547801  0.15961004   0   0
#>   -0.33296007  0.16830424   0   0
#>   -4.57211085 -0.49902927   0   1
#>    0.27602811  0.11925190   0   0
#>    0.32331468  0.15537921   0   0
#>    0.34966421  0.21729285   0   0
#>    0.34769988  0.22666532   0   0
#>    0.51885694  0.84647861   0   0
#>    0.45977445  0.80813040   0   0
#>    0.45004006  0.73972948   0   0
#>    0.51273435  0.70209437   0   0
#>    0.55281891  0.74032518   0   0
#>    0.58389922  0.82354596   0   0
#>    0.56290367  0.85388141   0   0
#>    0.56033486  0.89802502   0   0
#>    0.68866336  0.99301689   0   0
#>    0.72025435  1.17536224   0   0
#>    0.46175301  4.42101852   0   1
#>    0.17584152  0.48475838   0   0
#>    0.25554419  0.44728945   0   0
#>    0.29130307  0.44288752   0   0
#>    0.31773023  0.53627637   0   0
#>    0.26994841  0.60896423   0   0
#>    0.24631727  0.62997900   0   0
#>    1.07146642  0.77419740   0   0
#>    1.01413428  0.74138499   0   0
#>   12.09538935  0.48132782   0   0
#>    0.16634883  0.34108457   0   0
#>    0.70267412  0.22833597   0   0
#>    0.76430747  0.33087016   0   0
#>    0.76149405  0.33525129   0   0
#>    0.85307561  0.73728036   0   0
#>    0.81848225  0.76181010   0   0
#>    0.76680571  0.73061948   0   0
#>    0.74167593  0.68090494   0   0
#>    0.88039550  0.64413502   0   0
#>    0.86723004  0.72174798   0   0
#>    0.85209288  0.16498606   0   0
#>    0.91247486  0.17502713   0   0
#>    0.92021805  0.19517862   0   0
#>    0.91074245  0.21878205   0   0
#>    0.80615220  0.23936746   0   0
#>    0.80185050  0.23142148   0   0
#>    0.80423496  0.15992307   0   0
#>    0.24379996  0.92857510   0   0
#>    0.83618519  0.83053462   0   0
#>    0.79846262  0.89292992   0   0
#>    0.57684930  0.19060900   0   0
#>    0.62336270  0.16002870   0   0
#>    0.67478133  0.14857982   0   0
#>    0.47501228  0.37380552   0   0
#>    0.46385483  0.39193313   0   0
#>    0.37748145  0.41261408   0   0
#>    0.32451399  0.40564151   0   0
#>    0.31682654  0.35351992   0   0
#>    0.34495349  0.27940447   0   0
#>    0.44131836  0.30888701   0   0
#>    0.67272413  0.79248977   0   0
#>    0.65812503  0.81928306   0   0
#>    0.11085491 -0.02081340   0   0
#>    0.43954215  0.58210283   0   0
#>    0.42204299  0.58568496   0   0
#>    0.36877081  0.91032283   0   0
#>    0.94502764  0.12165597   0   0
#>    0.96023541  0.06710227   0   0
#>    1.10742875  0.10126946   0   0
#>    1.45557497  0.31386176   0   0
#>    0.91582798  0.34777349   0   0
#>    0.92749694  0.29976569   0   0
#>    0.52465295  0.12723164   0   0
#>    0.48233943  0.10732606   0   0
#>    0.45588384  0.06753272   0   0
#>    0.46996967  0.22518297   0   0
#>    0.46228836  0.16008493   0   0
#>    0.48139700  0.22168810   0   0
#>    0.55114364  0.50738596   0   0
#>    0.70480680  0.62356147   0   0
#>   35.80340781  0.39599135   0   1
#>    0.92076755  0.14365013   0   0
#>    0.49097946  0.22286465   0   0
#>    0.51113345  0.19858095   0   0
#>    0.55628938  0.36812143   0   0
#>    0.85676179 -0.63171977   0   0
#>    0.67981329  0.13608692   0   0
#>    0.58493466  0.07491615   0   0
#>    0.95568833 -2.12831987   0   0
#>    0.81334448  0.58110933   0   0
#>    0.88701137  0.44705668   0   0
#>    0.82315925  0.32693032   0   0
#>   -4.79123069  0.41075607   0   1
#>    0.66741107 -3.71148607   0   1
#>  -17.01315028 -7.41074224   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.11548870 0.24875967
#>   [2,] 0.00000000 0.75140262
#>   [3,] 0.63715682 0.46423124
#>   [4,] 0.62752783 0.48930559
#>   [5,] 0.85453869 0.95075594
#>   [6,] 0.23073568 0.00000000
#>   [7,] 0.62304398 0.29235869
#>   [8,] 0.95736076 0.61605692
#>   [9,] 0.09993590 0.65693739
#>  [10,] 0.01277186 0.80395800
#>  [11,] 0.51789052 0.02943382
#>  [12,] 0.11808146 0.18197979
#>  [13,] 0.74053398 0.86561653
#>  [14,] 0.04781754 0.11029868
#>  [15,] 0.25530327 0.24252237
#>  [16,] 0.49583010 0.79578547
#>  [17,] 0.58211118 0.97561563
#>  [18,] 0.24187591 0.54670249
#>  [19,] 0.99325594 0.64925890
#>  [20,] 0.17027207 0.39438917
#>  [21,] 0.69285214 0.29490688
#>  [22,] 0.81386719 0.66957112
#>  [23,] 0.83771589 0.21210868
#>  [24,] 0.08008110 0.18399115
#>  [25,] 0.25440666 0.71298060
#>  [26,] 0.11696131 0.32064823
#>  [27,] 0.76469743 0.85572093
#>  [28,] 0.58446970 0.21304931
#>  [29,] 0.41146320 0.32565091
#>  [30,] 0.07316262 0.52576207
#>  [31,] 0.63221084 0.74485160
#>  [32,] 0.16543167 0.05309614
#>  [33,] 0.38780006 0.50540891
#>  [34,] 0.48229968 0.92258788
#>  [35,] 0.92412569 0.77004814
#>  [36,] 0.99086784 0.09024212
#>  [37,] 0.93481422 0.69008732
#>  [38,] 0.96617058 0.33719158
#>  [39,] 0.51947398 0.08654457
#>  [40,] 0.07993559 0.08539462
#>  [41,] 0.43326658 0.16678382
#>  [42,] 0.56799927 0.54610068
#>  [43,] 0.68974201 0.73231408
#>  [44,] 1.00000000 0.90311232
#>  [45,] 0.44880701 0.13352571
#>  [46,] 0.96811211 0.26445626
#>  [47,] 0.85374017 0.11574659
#>  [48,] 0.01563722 0.47624928
#>  [49,] 0.42170633 0.80633430
#>  [50,] 0.09275120 0.44879157
#>  [51,] 0.56357792 0.21024336
#>  [52,] 0.59484220 0.38441551
#>  [53,] 0.99282890 0.15250663
#>  [54,] 0.35416351 0.74306667
#>  [55,] 0.38951596 0.06685309
#>  [56,] 0.60552073 0.68077617
#>  [57,] 0.47221085 0.29412174
#>  [58,] 0.68661948 0.36232864
#>  [59,] 0.21817377 0.35415930
#>  [60,] 0.84569162 0.25263175
#>  [61,] 0.92204966 0.07105789
#>  [62,] 0.62579750 0.03522434
#>  [63,] 0.46318040 0.84608837
#>  [64,] 0.33823762 0.61004647
#>  [65,] 0.48052363 0.88685914
#>  [66,] 0.76076275 0.52909979
#>  [67,] 0.90399998 0.79668196
#>  [68,] 0.44716445 0.47475611
#>  [69,] 0.42209811 0.67295935
#>  [70,] 0.58261978 0.40204030
#>  [71,] 0.55230052 0.54959501
#>  [72,] 0.26450737 0.26988091
#>  [73,] 0.84710004 0.90553939
#>  [74,] 0.88827318 1.00000000
#>  [75,] 0.49000661 0.14918367
#>  [76,] 0.81000278 0.65499210
#>  [77,] 0.51350895 0.52653202
#>  [78,] 0.55931290 0.02403652
#>  [79,] 0.54663249 0.40884686
#>  [80,] 0.82673930 0.38088506
#>  [81,] 0.86587913 0.36008062
#>  [82,] 0.17707763 0.04287172
#>  [83,] 0.99792897 0.05982255
#>  [84,] 0.64090254 0.89619192
#>  [85,] 0.32850499 0.64411696
#>  [86,] 0.55883075 0.32634395
#>  [87,] 0.73471787 0.80070554
#>  [88,] 0.06938922 0.06117343
#>  [89,] 0.05012976 0.20828447
#>  [90,] 0.87130884 0.53517499
#>  [91,] 0.84789420 0.11589828
#>  [92,] 0.78315185 0.94078571
#>  [93,] 0.76607184 0.25089448
#>  [94,] 0.56564811 0.12851876
#>  [95,] 0.55926828 0.17471732
#>  [96,] 0.99721005 0.14076486
#>  [97,] 0.01158040 0.34049048
#>  [98,] 0.76735153 0.20976205
#>  [99,] 0.42595367 0.27828823
#> [100,] 0.87595923 0.31526447
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.1154887,0.2487597) [8]: 12 15 24 26 59 72 89 97 
#> 2: (0,0.7514026) [5]: 9 10 30 48 97 
#> 3: (0.6371568,0.4642312) [5]: 4 52 58 66 70 
#> 4: (0.6275278,0.4893056) [7]: 3 42 56 66 70 77 79 
#> 5: (0.8545387,0.9507559) [4]: 44 73 74 92 
#> 6: (0.2307357,0) [5]: 11 55 78 82 88 
#> 7: (0.623044,0.2923587) [5]: 21 28 52 58 86 
#> 8: (0.9573608,0.6160569) [5]: 19 37 38 76 90 
#> 9: (0.0999359,0.6569374) [5]: 2 10 18 25 30 
#> 10: (0.01277186,0.803958) [6]: 2 9 17 25 34 49 
#> 11: (0.5178905,0.02943382) [4]: 6 39 55 78 
#> 12: (0.1180815,0.1819798) [5]: 1 15 24 32 40 
#> 13: (0.740534,0.8656165) [4]: 27 84 87 92 
#> 14: (0.04781754,0.1102987) [5]: 24 40 88 89 97 
#> 15: (0.2553033,0.2425224) [8]: 1 12 32 41 55 72 82 99 
#> 16: (0.4958301,0.7957855) [7]: 31 49 56 63 65 69 84 
#> 17: (0.5821112,0.9756156) [6]: 10 34 65 74 84 92 
#> 18: (0.2418759,0.5467025) [9]: 9 20 25 30 33 50 59 64 85 
#> 19: (0.9932559,0.6492589) [6]: 8 35 37 38 44 96 
#> 20: (0.1702721,0.3943892) [4]: 18 26 50 59 
#> 21: (0.6928521,0.2949069) [6]: 7 28 58 80 93 98 
#> 22: (0.8138672,0.6695711) [6]: 35 37 43 67 76 87 
#> 23: (0.8377159,0.2121087) [7]: 46 47 53 60 91 93 98 
#> 24: (0.0800811,0.1839912) [5]: 1 12 14 40 89 
#> 25: (0.2544067,0.7129806) [6]: 9 10 18 49 54 85 
#> 26: (0.1169613,0.3206482) [5]: 1 20 50 59 97 
#> 27: (0.7646974,0.8557209) [5]: 13 67 73 87 92 
#> 28: (0.5844697,0.2130493) [7]: 7 21 51 86 94 95 98 
#> 29: (0.4114632,0.3256509) [7]: 33 57 59 68 72 79 99 
#> 30: (0.07316262,0.5257621) [5]: 2 9 18 48 50 
#> 31: (0.6322108,0.7448516) [5]: 16 43 56 84 87 
#> 32: (0.1654317,0.05309614) [5]: 12 15 40 82 88 
#> 33: (0.3878001,0.5054089) [7]: 18 29 59 64 68 69 77 
#> 34: (0.4822997,0.9225879) [5]: 10 17 49 63 65 
#> 35: (0.9241257,0.7700481) [5]: 19 22 37 44 67 
#> 36: (0.9908678,0.09024212) [3]: 61 83 96 
#> 37: (0.9348142,0.6900873) [5]: 8 19 22 35 76 
#> 38: (0.9661706,0.3371916) [7]: 8 19 46 81 90 96 100 
#> 39: (0.519474,0.08654457) [6]: 11 45 55 75 78 94 
#> 40: (0.07993559,0.08539462) [5]: 12 14 24 32 88 
#> 41: (0.4332666,0.1667838) [6]: 15 45 55 57 75 99 
#> 42: (0.5679993,0.5461007) [4]: 4 56 71 77 
#> 43: (0.689742,0.7323141) [6]: 22 31 56 66 76 87 
#> 44: (1,0.9031123) [8]: 5 19 35 67 73 74 83 96 
#> 45: (0.448807,0.1335257) [4]: 39 41 55 75 
#> 46: (0.9681121,0.2644563) [6]: 23 38 53 60 96 100 
#> 47: (0.8537402,0.1157466) [4]: 23 53 61 91 
#> 48: (0.01563722,0.4762493) [4]: 2 30 50 97 
#> 49: (0.4217063,0.8063343) [7]: 10 16 25 34 54 63 69 
#> 50: (0.0927512,0.4487916) [6]: 18 20 26 30 48 97 
#> 51: (0.5635779,0.2102434) [5]: 28 57 75 86 95 
#> 52: (0.5948422,0.3844155) [6]: 3 7 58 70 79 86 
#> 53: (0.9928289,0.1525066) [5]: 23 46 47 61 96 
#> 54: (0.3541635,0.7430667) [4]: 25 49 69 85 
#> 55: (0.389516,0.06685309) [7]: 6 11 15 39 41 45 82 
#> 56: (0.6055207,0.6807762) [8]: 4 16 31 42 43 66 69 71 
#> 57: (0.4722109,0.2941217) [7]: 29 41 51 75 79 86 99 
#> 58: (0.6866195,0.3623286) [6]: 3 7 21 52 66 80 
#> 59: (0.2181738,0.3541593) [7]: 1 18 20 26 29 33 72 
#> 60: (0.8456916,0.2526318) [4]: 23 46 93 100 
#> 61: (0.9220497,0.07105789) [7]: 36 47 53 62 83 91 96 
#> 62: (0.6257975,0.03522434) [6]: 61 78 83 91 94 98 
#> 63: (0.4631804,0.8460884) [4]: 16 34 49 65 
#> 64: (0.3382376,0.6100465) [4]: 18 33 69 85 
#> 65: (0.4805236,0.8868591) [5]: 16 17 34 63 84 
#> 66: (0.7607628,0.5290998) [8]: 3 4 43 56 58 76 80 90 
#> 67: (0.904,0.796682) [6]: 22 27 35 44 73 87 
#> 68: (0.4471645,0.4747561) [4]: 29 33 77 79 
#> 69: (0.4220981,0.6729594) [9]: 16 33 49 54 56 64 71 77 85 
#> 70: (0.5826198,0.4020403) [4]: 3 4 52 79 
#> 71: (0.5523005,0.549595) [4]: 42 56 69 77 
#> 72: (0.2645074,0.2698809) [5]: 1 15 29 59 99 
#> 73: (0.8471,0.9055394) [5]: 5 27 44 67 92 
#> 74: (0.8882732,1) [4]: 5 17 44 92 
#> 75: (0.4900066,0.1491837) [7]: 39 41 45 51 57 94 95 
#> 76: (0.8100028,0.6549921) [6]: 8 22 37 43 66 90 
#> 77: (0.5135089,0.526532) [7]: 4 33 42 68 69 71 79 
#> 78: (0.5593129,0.02403652) [6]: 6 11 39 62 83 94 
#> 79: (0.5466325,0.4088469) [8]: 4 29 52 57 68 70 77 86 
#> 80: (0.8267393,0.3808851) [7]: 21 58 66 81 90 93 100 
#> 81: (0.8658791,0.3600806) [4]: 38 80 90 100 
#> 82: (0.1770776,0.04287172) [5]: 6 15 32 55 88 
#> 83: (0.997929,0.05982255) [6]: 36 44 61 62 78 96 
#> 84: (0.6409025,0.8961919) [7]: 13 16 17 31 65 87 92 
#> 85: (0.328505,0.644117) [5]: 18 25 54 64 69 
#> 86: (0.5588307,0.3263439) [6]: 7 28 51 52 57 79 
#> 87: (0.7347179,0.8007055) [7]: 13 22 27 31 43 67 84 
#> 88: (0.06938922,0.06117343) [5]: 6 14 32 40 82 
#> 89: (0.05012976,0.2082845) [4]: 1 14 24 97 
#> 90: (0.8713088,0.535175) [6]: 8 38 66 76 80 81 
#> 91: (0.8478942,0.1158983) [5]: 23 47 61 62 98 
#> 92: (0.7831519,0.9407857) [7]: 5 13 17 27 73 74 84 
#> 93: (0.7660718,0.2508945) [6]: 21 23 60 80 98 100 
#> 94: (0.5656481,0.1285188) [7]: 28 39 62 75 78 95 98 
#> 95: (0.5592683,0.1747173) [4]: 28 51 75 94 
#> 96: (0.99721,0.1407649) [8]: 19 36 38 44 46 53 61 83 
#> 97: (0.0115804,0.3404905) [7]: 1 2 14 26 48 50 89 
#> 98: (0.7673515,0.209762) [7]: 21 23 28 62 91 93 94 
#> 99: (0.4259537,0.2782882) [5]: 15 29 41 57 72 
#> 100: (0.8759592,0.3152645) [6]: 38 46 60 80 81 93 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  2 6 10 14 17 44 74 78 83 88 97 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.137711585 0.038564856 0.038053560 0.149984093 0.028865295 0.096175099 
#>                                                                         
#> 0.106615917 0.107232693 0.135819409 0.075899188 0.133879699 0.170917027 
#>                                                                         
#> 0.021448091 0.094397563 0.182509835 0.139050800 0.098142567 0.217038728 
#>                                                                         
#> 0.139804670 0.057897565 0.122257778 0.036625320 0.120651669 0.073018311 
#>                                                                         
#> 0.152919226 0.147885444 0.041772531 0.133716972 0.213787635 0.109232532 
#>                                                                         
#> 0.068331036 0.116159506 0.135821265 0.067709232 0.134590392 0.219536282 
#>                                                                         
#> 0.313240526 0.071290916 0.129651732 0.170007870 0.166749595 0.258194234 
#>                                                                         
#> 0.191583127 0.131097099 0.113023351 0.074124187 0.131441395 0.098596644 
#>                                                                         
#> 0.091393493 0.209664498 0.149075390 0.133861732 0.062653549 0.087043569 
#>                                                                         
#> 0.026417646 0.015497330 0.109088346 0.111743958 0.161665051 0.160650235 
#>                                                                         
#> 0.044306243 0.057449843 0.133375517 0.183589100 0.069411985 0.151589703 
#>                                                                         
#> 0.035772855 0.079423990 0.094705363 0.062554099 0.227265314 0.115781930 
#>                                                                         
#> 0.035433349 0.101157125 0.762352568 0.153176330 0.080672039 0.198864036 
#>                                                                         
#> 0.130410290 0.182911375 0.041152338 0.143450652 0.141343207 0.033382697 
#>                                                                         
#> 0.169329922 0.081854878 0.142973219 0.123683174 0.005847947 0.143864430 
#>                                                                         
#> 0.114645711 0.122990225 0.079638731 0.102222331 0.092546339 0.135343360 
#>                                                                         
#> 0.122831363 0.116197597 0.035786488 0.046636997 0.217450548 0.048891995 
#>                                                                         
#> 0.136749533 0.119322757 0.066811114 0.081856529 0.098738948 0.084156617 
#>                                                                         
#> 0.172617803 0.104836122 0.143329783 0.080945504 0.102509241 0.076706598 
#>                                                                         
#> 0.105231304 0.127067906 0.175178536 0.045935782 0.104674127 0.078413487 
#>                                                                         
#> 0.044074693 0.089222570 0.146026486 0.141565840 0.045129674 0.135179337 
#>                                                                         
#> 0.069562814 0.012532504 0.086658487 0.107905502 0.097622465 0.092419031 
#>                                                                         
#> 0.073818330 0.083399812 0.097936490 0.179363958 0.162235975 0.127352790 
#>                                                                         
#> 0.236294784 0.372943560 0.440073512 0.160598340 0.076876738 0.073815033 
#>                                                                         
#> 0.066830198 0.139953627 0.150508051 0.147150621 0.071903642 0.138605465 
#>                                                                         
#> 0.275597322 0.237205461 0.137516784 0.054085002 0.090338684 0.113272649 
#>                                                                         
#> 0.139593537 0.026859648 0.082716395 0.192731147 0.082275945 0.119942115 
#>                                                                         
#> 0.114094962 0.072079721 0.045824342 0.153065044 0.059690812 0.288659401 
#>                                                                         
#> 0.172280331 0.068681667 0.172553964 0.096279795 0.072652162 0.088192700 
#>                                                                         
#> 0.069854655 0.094539094 0.118096658 0.279004469 0.048896167 0.077387631 
#>                                                                         
#> 0.152414970 0.179718383 0.164323019 0.408941378 0.484282384 0.594654193 
#>                                                                         
#> 0.057132691 0.103845091 0.137306357 0.104217454 0.065171031 0.026111212 
#>                                                                         
#> 0.086410006 0.040642115 0.080445746 0.098013070 0.233026596 0.214428624 
#>                                                                         
#> 0.221071812 0.193409588 0.174358127 0.092350966 0.059969864 0.074870627 
#>                                                                         
#> 0.143257325 0.158931349 0.145581471 0.176429784 0.134899024 0.098815734 
#>                                                                         
#> 0.204035471 0.307131512 0.178394869 0.168304660 0.193996581 0.151654271 
#>                                                                         
#> 0.115317126 0.130361712 0.253942994 0.139172677 0.508509410 0.090993207 
#>                                                                         
#> 0.113136211 0.085429599 0.159116365 0.149172945 0.155823889 0.153169358 
#>                                                                         
#> 0.139081773 0.015082491 0.122674772 0.096747295 0.097685365 0.166169944 
#>                                                                         
#> 0.140511368 0.041300504 0.081469059 0.070403479 0.104195021 0.151297057 
#>                                                                         
#> 0.096291576 0.021079367 0.045874316 0.086600613 0.068442424 0.158720842 
#>                                                                         
#> 0.153319738 0.181308803 0.195380484 0.157182401 0.049529759 0.058881446 
#>                                                                         
#> 0.116736256 0.096381505 0.127471216 0.171024879 0.078852532 0.050919253 
#>                                                                         
#> 0.071442115 0.031228356 0.072761226 0.102870224 0.092837952 0.062400988 
#>                                                                         
#> 0.069224150 0.084858990 0.133160067 0.036709766 0.059407043 0.016082946 
#>                                                                         
#> 0.215267258 0.843292317 0.081628828 0.124030333 0.095608702 0.054046892 
#>                                                                         
#> 0.298411435 0.224724169 0.111003593 0.067419359 0.110712896 0.044325513 
#>                                                 
#> 0.082028434 0.411075297 0.329455218 0.053652844 
#> 
#> $alpha
#> [1] 0.2711053
#> 
#> attr(,"class")
#> [1] "scree"