Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    20   82 0.091417905 0.77806408 0.099359264 0.84091409  0.086040474
#>    20   87 0.091417905 0.77806408 0.099657885 0.74995152  0.187019548
#>    30   87 0.032388486 0.86739888 0.099657885 0.74995152 -0.124634166
#>     1   30 0.070143781 0.83964222 0.032388486 0.86739888 -0.002800944
#>    68   30 0.070505823 0.91851314 0.032388486 0.86739888  0.070055944
#>    62   97 0.105457463 0.70847935 0.127190263 0.66622893  0.177099467
#>    36   62 0.104036089 0.67461934 0.105457463 0.70847935 -0.104097662
#>    82   68 0.099359264 0.84091409 0.070505823 0.91851314  0.083064969
#>    30   36 0.032388486 0.86739888 0.104036089 0.67461934 -0.125316649
#>    97  100 0.127190263 0.66622893 0.154874862 0.64551367  0.180998220
#>    87   62 0.099657885 0.74995152 0.105457463 0.70847935  0.175822781
#>    55  100 0.156119144 0.61600862 0.154874862 0.64551367  0.134470507
#>    94   74 0.249443024 0.71729795 0.286456860 0.88635535  0.204548254
#>    55   33 0.156119144 0.61600862 0.213829828 0.57049341  0.216589895
#>    36   52 0.104036089 0.67461934 0.132884553 0.60421494  0.101921384
#>    74   68 0.286456860 0.88635535 0.070505823 0.91851314  0.180383021
#>    33   92 0.213829828 0.57049341 0.220433434 0.52630905  0.236286488
#>    33   94 0.213829828 0.57049341 0.249443024 0.71729795  0.230405204
#>    55   52 0.156119144 0.61600862 0.132884553 0.60421494  0.165557154
#>    94   58 0.249443024 0.71729795 0.265850059 0.55100755  0.272524824
#>    33   52 0.213829828 0.57049341 0.132884553 0.60421494  0.153145570
#>    50   28 0.349026121 0.59885808 0.480508405 0.76738333  0.370290518
#>    74   49 0.286456860 0.88635535 0.323451188 0.85017619  0.361036766
#>    37   72 0.467790799 0.84408077 0.445291637 0.92352051  0.396450944
#>    37   49 0.467790799 0.84408077 0.323451188 0.85017619  0.393306744
#>    54   63 0.283732973 0.47694649 0.368763747 0.53386601  0.314149495
#>    58   92 0.265850059 0.55100755 0.220433434 0.52630905  0.258679622
#>    39   28 0.573821253 0.73025251 0.480508405 0.76738333  0.512814786
#>    63   50 0.368763747 0.53386601 0.349026121 0.59885808  0.319288508
#>    28   49 0.480508405 0.76738333 0.323451188 0.85017619  0.365552827
#>    92   42 0.220433434 0.52630905 0.053190139 0.41324929  0.151237684
#>    92   52 0.220433434 0.52630905 0.132884553 0.60421494  0.117333120
#>    56   51 0.487981730 0.83209176 0.507162427 0.89565174  0.493862008
#>    37   28 0.467790799 0.84408077 0.480508405 0.76738333  0.456984927
#>    51   72 0.507162427 0.89565174 0.445291637 0.92352051  0.465792572
#>    63   13 0.368763747 0.53386601 0.351100769 0.43875840  0.413901795
#>    54   15 0.283732973 0.47694649 0.201543938 0.35722037  0.218501657
#>    54   92 0.283732973 0.47694649 0.220433434 0.52630905  0.205848313
#>    13   54 0.351100769 0.43875840 0.283732973 0.47694649  0.297996962
#>    56   28 0.487981730 0.83209176 0.480508405 0.76738333  0.537166448
#>    63   19 0.368763747 0.53386601 0.595711691 0.59374733  0.472463154
#>    15   88 0.201543938 0.35722037 0.080814051 0.30616516  0.189435848
#>    15   42 0.201543938 0.35722037 0.053190139 0.41324929  0.123201248
#>    56    3 0.487981730 0.83209176 0.599582736 0.79189514  0.549342661
#>     3   39 0.599582736 0.79189514 0.573821253 0.73025251  0.611568599
#>    70   67 0.595488716 0.92977333 0.647820214 0.96154275  0.512084968
#>    90   70 0.610027083 0.83764329 0.595488716 0.92977333  0.564841696
#>    63   75 0.368763747 0.53386601 0.484345876 0.45653669  0.479470738
#>     8   75 0.390131261 0.30561636 0.484345876 0.45653669  0.425639467
#>    39   66 0.573821253 0.73025251 0.576280726 0.71259942  0.621730753
#>    66   19 0.576280726 0.71259942 0.595711691 0.59374733  0.644081519
#>    90   18 0.610027083 0.83764329 0.651974202 0.76515920  0.681886369
#>    67   68 0.647820214 0.96154275 0.070505823 0.91851314  0.235472581
#>    90   99 0.610027083 0.83764329 0.663711129 0.90070319  0.617131227
#>    99   67 0.663711129 0.90070319 0.647820214 0.96154275  0.633939327
#>    98   75 0.639866695 0.53272575 0.484345876 0.45653669  0.568979058
#>     7    8 0.270893179 0.37547542 0.390131261 0.30561636  0.344578328
#>    15   79 0.201543938 0.35722037 0.349692642 0.24001811  0.259432618
#>    73    6 0.705394865 0.95260736 0.673577916 0.93020614  0.697401483
#>    18   40 0.651974202 0.76515920 0.658950399 0.57979117  0.717856298
#>    18   66 0.651974202 0.76515920 0.576280726 0.71259942  0.660115344
#>    98   40 0.639866695 0.53272575 0.658950399 0.57979117  0.622963971
#>     8   79 0.390131261 0.30561636 0.349692642 0.24001811  0.401213183
#>    67    6 0.647820214 0.96154275 0.673577916 0.93020614  0.676941198
#>     6   99 0.673577916 0.93020614 0.663711129 0.90070319  0.699559257
#>    18   46 0.651974202 0.76515920 0.749136154 0.86191306  0.683449598
#>    77   43 0.603334583 0.39959835 0.611324414 0.33014148  0.721223409
#>     5   77 0.518065435 0.27019411 0.603334583 0.39959835  0.513985667
#>    88   14 0.080814051 0.30616516 0.003008577 0.15346237  0.108966443
#>    88   42 0.080814051 0.30616516 0.053190139 0.41324929 -0.118900463
#>    42   30 0.053190139 0.41324929 0.032388486 0.86739888 -1.462876966
#>    46   38 0.749136154 0.86191306 0.720867015 0.93427972  0.709528438
#>    35   18 0.816900638 0.75454600 0.651974202 0.76515920  0.729473419
#>    16   84 0.766635931 0.90105883 0.784232714 1.00000000  0.768498071
#>    98   77 0.639866695 0.53272575 0.603334583 0.39959835  0.675526826
#>    41   65 0.140786074 0.08612974 0.302313722 0.11877572  0.198701472
#>    14   30 0.003008577 0.15346237 0.032388486 0.86739888 -4.990336856
#>     8   17 0.390131261 0.30561636 0.410980602 0.20116389  0.447115014
#>    85   41 0.094660769 0.05640766 0.140786074 0.08612974  0.078175507
#>     8    5 0.390131261 0.30561636 0.518065435 0.27019411  0.473608999
#>     5   17 0.518065435 0.27019411 0.410980602 0.20116389  0.450631818
#>    65   86 0.302313722 0.11877572 0.422863708 0.19442406  0.378855452
#>    47   16 0.823380055 0.80599655 0.766635931 0.90105883  0.805892747
#>     2   43 0.542551234 0.24072417 0.611324414 0.33014148  0.571456866
#>    67   73 0.647820214 0.96154275 0.705394865 0.95260736  0.692974996
#>    84   73 0.784232714 1.00000000 0.705394865 0.95260736  0.747251701
#>    17   79 0.410980602 0.20116389 0.349692642 0.24001811  0.348618466
#>    65   79 0.302313722 0.11877572 0.349692642 0.24001811  0.215471024
#>    27   40 0.753835840 0.49330545 0.658950399 0.57979117  0.704133368
#>    86    2 0.422863708 0.19442406 0.542551234 0.24072417  0.483194977
#>    81   26 0.270574255 0.05295177 0.319237660 0.04393118  0.297686496
#>    32   84 0.938375320 0.97713363 0.784232714 1.00000000  0.853599085
#>    35   47 0.816900638 0.75454600 0.823380055 0.80599655  0.752264414
#>    27   59 0.753835840 0.49330545 0.777539212 0.58027723  0.718369186
#>    57   48 0.622209545 0.22632443 0.690093341 0.25756254  0.643905482
#>    26   65 0.319237660 0.04393118 0.302313722 0.11877572  0.319447620
#>    34   85 0.000000000 0.12739310 0.094660769 0.05640766  0.077172183
#>     2   25 0.542551234 0.24072417 0.610342905 0.23497386  0.580266143
#>    57   25 0.622209545 0.22632443 0.610342905 0.23497386  0.589522647
#>    26    9 0.319237660 0.04393118 0.466931547 0.05231843  0.390178011
#>    41   11 0.140786074 0.08612974 0.234709979 0.03029688  0.210457757
#>    47   32 0.823380055 0.80599655 0.938375320 0.97713363  0.870459285
#>    29   24 0.883498795 0.64071488 0.903812483 0.74080726  0.852674277
#>    35   59 0.816900638 0.75454600 0.777539212 0.58027723  0.798178809
#>    48   43 0.690093341 0.25756254 0.611324414 0.33014148  0.640024691
#>    86   93 0.422863708 0.19442406 0.515410900 0.08347058  0.502393034
#>    11   26 0.234709979 0.03029688 0.319237660 0.04393118  0.284340446
#>    11   81 0.234709979 0.03029688 0.270574255 0.05295177  0.225907045
#>     9   86 0.466931547 0.05231843 0.422863708 0.19442406  0.393451912
#>    11   85 0.234709979 0.03029688 0.094660769 0.05640766  0.132018371
#>    27   91 0.753835840 0.49330545 0.917220044 0.59831572  0.857387675
#>    27   43 0.753835840 0.49330545 0.611324414 0.33014148  0.725098920
#>    57   80 0.622209545 0.22632443 0.602433823 0.14978332  0.654086168
#>    24   23 0.903812483 0.74080726 1.000000000 0.96129331  0.984930971
#>    24   61 0.903812483 0.74080726 0.876678606 0.77303376  0.857453263
#>    93   80 0.515410900 0.08347058 0.602433823 0.14978332  0.523560320
#>     9   93 0.466931547 0.05231843 0.515410900 0.08347058  0.453757201
#>    44    9 0.454586314 0.00000000 0.466931547 0.05231843  0.393429782
#>    29   91 0.883498795 0.64071488 0.917220044 0.59831572  0.910507171
#>    47   61 0.823380055 0.80599655 0.876678606 0.77303376  0.903670627
#>    23   32 1.000000000 0.96129331 0.938375320 0.97713363  0.942290874
#>    80   21 0.602433823 0.14978332 0.592714852 0.10314611  0.654681996
#>    22   71 0.866091855 0.39847689 0.965549671 0.56659466  0.886995564
#>    27   48 0.753835840 0.49330545 0.690093341 0.25756254  0.745149290
#>    11   44 0.234709979 0.03029688 0.454586314 0.00000000  0.315102944
#>    11   96 0.234709979 0.03029688 0.003129441 0.01036572  0.205979838
#>    93   76 0.515410900 0.08347058 0.490663101 0.01629187  0.529833057
#>    48   69 0.690093341 0.25756254 0.706251417 0.15076989  0.760872190
#>    21    4 0.592714852 0.10314611 0.572632413 0.02372969  0.661580007
#>    71   23 0.965549671 0.56659466 1.000000000 0.96129331  1.307698857
#>    71   95 0.965549671 0.56659466 0.939158912 0.63658027  0.954464591
#>    21   93 0.592714852 0.10314611 0.515410900 0.08347058  0.560220413
#>    76   44 0.490663101 0.01629187 0.454586314 0.00000000  0.547836982
#>    24   95 0.903812483 0.74080726 0.939158912 0.63658027  1.147990397
#>    22   48 0.866091855 0.39847689 0.690093341 0.25756254  0.776699551
#>    48   60 0.690093341 0.25756254 0.725735753 0.13813381  0.794773998
#>    69   21 0.706251417 0.15076989 0.592714852 0.10314611  0.671149674
#>    45   71 0.982831856 0.31672730 0.965549671 0.56659466  0.983790793
#>    31   45 0.882338803 0.29034642 0.982831856 0.31672730  0.920047323
#>    31   60 0.882338803 0.29034642 0.725735753 0.13813381  0.826125270
#>    60   83 0.725735753 0.13813381 0.752038480 0.03256705  0.718904187
#>    60   21 0.725735753 0.13813381 0.592714852 0.10314611  0.672835922
#>    12   83 0.816843211 0.02638720 0.752038480 0.03256705  0.703882571
#>    60   89 0.725735753 0.13813381 0.756399338 0.03590510  0.797432559
#>    64   31 0.977919809 0.28392489 0.882338803 0.29034642  0.923481977
#>    12   53 0.816843211 0.02638720 0.875472645 0.08908238  0.797912488
#>    44   12 0.454586314 0.00000000 0.816843211 0.02638720  0.705255993
#>    23   45 1.000000000 0.96129331 0.982831856 0.31672730  9.349901888
#>    45   64 0.982831856 0.31672730 0.977919809 0.28392489  1.657360372
#>    64   53 0.977919809 0.28392489 0.875472645 0.08908238  0.971539525
#>     1   82 0.070143781 0.83964222 0.099359264 0.84091409  0.086040474
#>    94   87 0.249443024 0.71729795 0.099657885 0.74995152  0.187019548
#>    20   30 0.091417905 0.77806408 0.032388486 0.86739888 -0.002800944
#>    94   97 0.249443024 0.71729795 0.127190263 0.66622893  0.177099467
#>     1   68 0.070143781 0.83964222 0.070505823 0.91851314  0.083064969
#>    87   36 0.099657885 0.74995152 0.104036089 0.67461934 -0.125316649
#>    94  100 0.249443024 0.71729795 0.154874862 0.64551367  0.180998220
#>    97   36 0.127190263 0.66622893 0.104036089 0.67461934  0.122990822
#>    94   62 0.249443024 0.71729795 0.105457463 0.70847935  0.175822781
#>    52  100 0.132884553 0.60421494 0.154874862 0.64551367  0.134470507
#>    82   74 0.099359264 0.84091409 0.286456860 0.88635535  0.204548254
#>    94   20 0.249443024 0.71729795 0.091417905 0.77806408  0.189619366
#>   100   33 0.154874862 0.64551367 0.213829828 0.57049341  0.216589895
#>    97   52 0.127190263 0.66622893 0.132884553 0.60421494  0.101921384
#>    49   94 0.323451188 0.85017619 0.249443024 0.71729795  0.244911777
#>    50   58 0.349026121 0.59885808 0.265850059 0.55100755  0.272524824
#>    58   33 0.265850059 0.55100755 0.213829828 0.57049341  0.267706944
#>    94   28 0.249443024 0.71729795 0.480508405 0.76738333  0.370290518
#>    72   74 0.445291637 0.92352051 0.286456860 0.88635535  0.251587796
#>    72   49 0.445291637 0.92352051 0.323451188 0.85017619  0.361036766
#>    58   63 0.265850059 0.55100755 0.368763747 0.53386601  0.314149495
#>    66   28 0.576280726 0.71259942 0.480508405 0.76738333  0.512814786
#>    42   52 0.053190139 0.41324929 0.132884553 0.60421494  0.117333120
#>    42   36 0.053190139 0.41324929 0.104036089 0.67461934 -0.052375788
#>    37   51 0.467790799 0.84408077 0.507162427 0.89565174  0.493862008
#>    75   13 0.484345876 0.45653669 0.351100769 0.43875840  0.413901795
#>     7   15 0.270893179 0.37547542 0.201543938 0.35722037  0.218501657
#>    15   92 0.201543938 0.35722037 0.220433434 0.52630905  0.205848313
#>     7   54 0.270893179 0.37547542 0.283732973 0.47694649  0.297996962
#>     3   28 0.599582736 0.79189514 0.480508405 0.76738333  0.537166448
#>    66   50 0.576280726 0.71259942 0.349026121 0.59885808  0.470839082
#>    50   19 0.349026121 0.59885808 0.595711691 0.59374733  0.472463154
#>    41   88 0.140786074 0.08612974 0.080814051 0.30616516  0.189435848
#>    90    3 0.610027083 0.83764329 0.599582736 0.79189514  0.549342661
#>    18   39 0.651974202 0.76515920 0.573821253 0.73025251  0.611568599
#>    72   67 0.445291637 0.92352051 0.647820214 0.96154275  0.512084968
#>    70   51 0.595488716 0.92977333 0.507162427 0.89565174  0.517303962
#>    90   56 0.610027083 0.83764329 0.487981730 0.83209176  0.548382548
#>    19   75 0.595711691 0.59374733 0.484345876 0.45653669  0.479470738
#>    40   19 0.658950399 0.57979117 0.595711691 0.59374733  0.644081519
#>    99   18 0.663711129 0.90070319 0.651974202 0.76515920  0.681886369
#>    72   68 0.445291637 0.92352051 0.070505823 0.91851314  0.235472581
#>    70   99 0.595488716 0.92977333 0.663711129 0.90070319  0.617131227
#>    77   75 0.603334583 0.39959835 0.484345876 0.45653669  0.568979058
#>    13    8 0.351100769 0.43875840 0.390131261 0.30561636  0.344578328
#>     7   79 0.270893179 0.37547542 0.349692642 0.24001811  0.259432618
#>    38    6 0.720867015 0.93427972 0.673577916 0.93020614  0.697401483
#>    59   40 0.777539212 0.58027723 0.658950399 0.57979117  0.717856298
#>    40   66 0.658950399 0.57979117 0.576280726 0.71259942  0.660115344
#>    19   98 0.595711691 0.59374733 0.639866695 0.53272575  0.551865799
#>    38   99 0.720867015 0.93427972 0.663711129 0.90070319  0.699559257
#>    99   46 0.663711129 0.90070319 0.749136154 0.86191306  0.683449598
#>    41   14 0.140786074 0.08612974 0.003008577 0.15346237  0.108966443
#>    14   42 0.003008577 0.15346237 0.053190139 0.41324929 -0.118900463
#>    59   18 0.777539212 0.58027723 0.651974202 0.76515920  0.729473419
#>    38   84 0.720867015 0.93427972 0.784232714 1.00000000  0.768498071
#>    16   46 0.766635931 0.90105883 0.749136154 0.86191306  0.727385127
#>    27   77 0.753835840 0.49330545 0.603334583 0.39959835  0.675526826
#>    15   65 0.201543938 0.35722037 0.302313722 0.11877572  0.198701472
#>    34   30 0.000000000 0.12739310 0.032388486 0.86739888 -4.990336856
#>    75    5 0.484345876 0.45653669 0.518065435 0.27019411  0.473608999
#>    86   17 0.422863708 0.19442406 0.410980602 0.20116389  0.450631818
#>    43    5 0.611324414 0.33014148 0.518065435 0.27019411  0.528903197
#>    84   67 0.784232714 1.00000000 0.647820214 0.96154275  0.226778020
#>    38   73 0.720867015 0.93427972 0.705394865 0.95260736  0.747251701
#>     5    2 0.518065435 0.27019411 0.542551234 0.24072417  0.483194977
#>    78   26 0.280207157 0.08667062 0.319237660 0.04393118  0.297686496
#>    46   47 0.749136154 0.86191306 0.823380055 0.80599655  0.752264414
#>    25   48 0.610342905 0.23497386 0.690093341 0.25756254  0.643905482
#>    78   65 0.280207157 0.08667062 0.302313722 0.11877572  0.319447620
#>    78   41 0.280207157 0.08667062 0.140786074 0.08612974  0.210216798
#>    14   85 0.003008577 0.15346237 0.094660769 0.05640766  0.077172183
#>    43   25 0.611324414 0.33014148 0.610342905 0.23497386  0.580266143
#>    80   25 0.602433823 0.14978332 0.610342905 0.23497386  0.589522647
#>    65    9 0.302313722 0.11877572 0.466931547 0.05231843  0.390178011
#>    78   11 0.280207157 0.08667062 0.234709979 0.03029688  0.210457757
#>    16   32 0.766635931 0.90105883 0.938375320 0.97713363  0.870459285
#>    61   35 0.876678606 0.77303376 0.816900638 0.75454600  0.842565794
#>    35   24 0.816900638 0.75454600 0.903812483 0.74080726  0.852674277
#>    29   59 0.883498795 0.64071488 0.777539212 0.58027723  0.798178809
#>     2   93 0.542551234 0.24072417 0.515410900 0.08347058  0.502393034
#>    78   81 0.280207157 0.08667062 0.270574255 0.05295177  0.225907045
#>    96   85 0.003129441 0.01036572 0.094660769 0.05640766  0.132018371
#>    96   34 0.003129441 0.01036572 0.000000000 0.12739310  0.030650374
#>    59   91 0.777539212 0.58027723 0.917220044 0.59831572  0.857387675
#>    22   27 0.866091855 0.39847689 0.753835840 0.49330545  0.860953370
#>    69   80 0.706251417 0.15076989 0.602433823 0.14978332  0.654086168
#>    61   23 0.876678606 0.77303376 1.000000000 0.96129331  0.984930971
#>    80    2 0.602433823 0.14978332 0.542551234 0.24072417  0.572854201
#>    95   91 0.939158912 0.63658027 0.917220044 0.59831572  0.910507171
#>    32   61 0.938375320 0.97713363 0.876678606 0.77303376  0.903670627
#>    69   57 0.706251417 0.15076989 0.622209545 0.22632443  0.675139119
#>    95   29 0.939158912 0.63658027 0.883498795 0.64071488  0.914880037
#>    91   71 0.917220044 0.59831572 0.965549671 0.56659466  0.886995564
#>    76    9 0.490663101 0.01629187 0.466931547 0.05231843  0.502656149
#>    26   44 0.319237660 0.04393118 0.454586314 0.00000000  0.315102944
#>    44   96 0.454586314 0.00000000 0.003129441 0.01036572  0.205979838
#>     4   76 0.572632413 0.02372969 0.490663101 0.01629187  0.529833057
#>    60   69 0.725735753 0.13813381 0.706251417 0.15076989  0.760872190
#>    83    4 0.752038480 0.03256705 0.572632413 0.02372969  0.661580007
#>    95   23 0.939158912 0.63658027 1.000000000 0.96129331  1.307698857
#>     4   93 0.572632413 0.02372969 0.515410900 0.08347058  0.560220413
#>     4   44 0.572632413 0.02372969 0.454586314 0.00000000  0.547836982
#>    31   48 0.882338803 0.29034642 0.690093341 0.25756254  0.776699551
#>    22   45 0.866091855 0.39847689 0.982831856 0.31672730  0.920047323
#>    53   60 0.875472645 0.08908238 0.725735753 0.13813381  0.826125270
#>    89   83 0.756399338 0.03590510 0.752038480 0.03256705  0.718904187
#>    83   21 0.752038480 0.03256705 0.592714852 0.10314611  0.672835922
#>    53   89 0.875472645 0.08908238 0.756399338 0.03590510  0.797432559
#>    53   31 0.875472645 0.08908238 0.882338803 0.29034642  0.923481977
#>    89   12 0.756399338 0.03590510 0.816843211 0.02638720  0.781496417
#>     4   12 0.572632413 0.02372969 0.816843211 0.02638720  0.705255993
#>    10   45 0.990561435 0.04325996 0.982831856 0.31672730  9.349901888
#>    10   64 0.990561435 0.04325996 0.977919809 0.28392489  1.657360372
#>    10   53 0.990561435 0.04325996 0.875472645 0.08908238  0.971539525
#>    10   12 0.990561435 0.04325996 0.816843211 0.02638720  0.907003541
#>     1   20 0.070143781 0.83964222 0.091417905 0.77806408  0.086040474
#>    82   94 0.099359264 0.84091409 0.249443024 0.71729795  0.204548254
#>    50   94 0.349026121 0.59885808 0.249443024 0.71729795  0.272524824
#>    58   54 0.265850059 0.55100755 0.283732973 0.47694649  0.314149495
#>    37   56 0.467790799 0.84408077 0.487981730 0.83209176  0.493862008
#>     7   13 0.270893179 0.37547542 0.351100769 0.43875840  0.297996962
#>    41   15 0.140786074 0.08612974 0.201543938 0.35722037  0.189435848
#>    18    3 0.651974202 0.76515920 0.599582736 0.79189514  0.611568599
#>    72   70 0.445291637 0.92352051 0.595488716 0.92977333  0.512084968
#>    51   90 0.507162427 0.89565174 0.610027083 0.83764329  0.564841696
#>    38   16 0.720867015 0.93427972 0.766635931 0.90105883  0.768498071
#>    27   98 0.753835840 0.49330545 0.639866695 0.53272575  0.675526826
#>    34   14 0.000000000 0.12739310 0.003008577 0.15346237 -4.990336856
#>    86    5 0.422863708 0.19442406 0.518065435 0.27019411  0.450631818
#>    17   65 0.410980602 0.20116389 0.302313722 0.11877572  0.378855452
#>    84   68 0.784232714 1.00000000 0.070505823 0.91851314  0.226778020
#>    35   46 0.816900638 0.75454600 0.749136154 0.86191306  0.735625670
#>    35   29 0.816900638 0.75454600 0.883498795 0.64071488  0.852674277
#>    22   91 0.866091855 0.39847689 0.917220044 0.59831572  0.860953370
#>    31   22 0.882338803 0.29034642 0.866091855 0.39847689  0.776699551
#>    10   23 0.990561435 0.04325996 1.000000000 0.96129331  9.349901888
#>            my1           mx2           my2 bp1 bp2
#>   0.8106702558   0.189619366  0.7975826335   0   0
#>   0.7908216857  -0.124634166  0.6994738874   0   0
#>   0.6994738874  -0.125316649  0.6990829869   0   0
#>   0.7799771677   0.070055944  0.8790789153   0   0
#>   0.8790789153 -12.502325030 10.2546551001   0   1
#>   0.7186159374   0.122990822  0.6907834943   0   0
#>   0.7003162055   0.122990822  0.6907834943   0   0
#>   0.8790191999   0.180383021  0.9152047021   0   0
#>   0.6990829869  -0.329270369  0.6232824374   0   0
#>   0.7092828171   0.125252744  0.6347825967   0   0
#>   0.7394610191  -0.104097662  0.7003162055   0   0
#>   0.6298744206   0.216589895  0.6333375438   0   0
#>   0.8157079654   0.244911777  0.8068706782   0   0
#>   0.6333375438   0.165557154  0.5686309471   0   0
#>   0.6326402488  -0.052375788  0.5694164030   0   0
#>   0.9152047021   0.251587796  1.3933701040   0   0
#>   0.5512640301   0.153145570  0.5388381436   0   0
#>   0.6441943579   0.267706944  0.6351453586   0   0
#>   0.5686309471   0.134470507  0.6298744206   0   0
#>   0.6356207146   0.267706944  0.6351453586   0   0
#>   0.5388381436   0.165557154  0.5686309471   0   0
#>   0.7178211664   0.470839082  0.6393738436   0   0
#>   0.9256121408   0.244911777  0.8068706782   0   0
#>   0.8667816936   0.465792572  0.8864208363   0   0
#>   0.7923270682   0.396450944  0.8667816936   0   0
#>   0.5234804690   0.336062449  0.4907452034   0   0
#>   0.5100866163   0.236286488  0.5512640301   0   0
#>   0.7127550651   0.539835411  0.7806601539   0   0
#>   0.5543338553   0.472463154  0.6008518955   0   0
#>   0.7396781691   0.393306744  0.7923270682   0   0
#>   0.4484397034   0.117333120  0.4985929248   0   0
#>   0.4985929248   0.153145570  0.5388381436   0   0
#>   0.8649913514   0.548382548  0.8485385127   0   0
#>   0.8028858834   0.393306744  0.7923270682   0   0
#>   0.8864208363   0.517303962  1.0007800203   0   0
#>   0.4762892124   0.336062449  0.4907452034   0   0
#>   0.4336527479   0.205848313  0.4423389564   0   0
#>   0.4423389564   0.258679622  0.5100866163   0   0
#>   0.4235936905   0.336062449  0.4907452034   0   0
#>   0.7936255362   0.456984927  0.8028858834   0   0
#>   0.6008518955   0.479470738  0.5742934213   0   0
#>   0.2175801213   0.123201248  0.3742046160   0   0
#>   0.3742046160   0.151237684  0.4484397034   0   0
#>   0.8274312983   0.537166448  0.7936255362   0   0
#>   0.7506816560   0.539835411  0.7806601539   0   0
#>   1.1261440159   0.633939327  0.9254220666   0   0
#>   0.8777250307   0.617131227  0.8859764590   0   0
#>   0.5742934213   0.413901795  0.4762892124   0   0
#>   0.3883174636   0.473608999  0.3583716585   0   0
#>   0.7279295028   0.512814786  0.7127550651   0   0
#>   0.6626696617   0.473164962  0.6347267475   0   0
#>   0.8308492207   0.640154764  0.8066988095   0   0
#>   2.5995425937   0.226778020  2.7161947075   0   0
#>   0.8859764590   0.681886369  0.8308492207   0   0
#>   0.9254220666   0.636691372  0.9261408835   0   0
#>   0.4806021854   0.551865799  0.5155346025   0   0
#>   0.3645544514   0.311814502  0.3086319207   0   0
#>   0.2781598534   0.215471024  0.2225905033   0   0
#>   0.9301647676   0.676941198  0.9592249657   0   0
#>   0.6748233408   0.660115344  0.6726503000   0   0
#>   0.6726503000   0.621730753  0.7279295028   0   0
#>   0.5669809950   0.704133368  0.5340690850   0   0
#>   0.2535213125   0.311814502  0.3086319207   0   0
#>   0.9592249657   0.636691372  0.9261408835   0   0
#>   0.9051157291   0.636691372  0.9261408835   0   0
#>   0.8307138587   0.735625670  0.7783177150   0   0
#>   0.3779714684   0.528903197  0.3558483032   0   0
#>   0.3656780048   0.528903197  0.3558483032   0   0
#>   0.1956476846  -0.118900463  0.3117509514   0   0
#>   0.3117509514   0.123201248  0.3742046160   0   0
#>   0.5713592555  -0.329270369  0.6232824374   0   0
#>   0.8881456193   0.727385127  0.8951211138   0   0
#>   0.6827132604   0.735625670  0.7783177150   0   0
#>   0.9517630346   0.853599085  0.9366277366   0   0
#>   0.4513639159   0.568979058  0.4806021854   0   0
#>   0.2155034745   0.210216798  0.1585272842   0   0
#>   0.7165212542  -1.462876966  0.5713592555   0   0
#>   0.2626835940   0.401213183  0.2535213125   0   0
#>   0.1326426055   0.158183609  0.0084790498   0   0
#>   0.3583716585   0.447115014  0.2626835940   0   0
#>   0.2572280646   0.447115014  0.2626835940   0   0
#>   0.1306779067   0.393451912  0.1074176070   0   0
#>   0.8600249653   0.870459285  0.8985657152   0   0
#>   0.2896483701   0.580266143  0.2828729256   0   0
#>   1.0625377188   0.676941198  0.9592249657   0   0
#>   0.9722482031   0.692974996  1.0625377188   0   0
#>   0.1705593471   0.401213183  0.2535213125   0   0
#>   0.2225905033   0.348618466  0.1705593471   0   0
#>   0.5340690850   0.718369186  0.5496875223   0   0
#>   0.2163138916   0.502393034  0.1666861925   0   0
#>   0.0634416660   0.284340446 -0.0085563660   0   0
#>   0.9366277366   1.795268118  7.2844342918   0   1
#>   0.7888192220   0.842565794  0.7774471310   0   0
#>   0.5496875223   0.857387675  0.5117992890   0   0
#>   0.2685552847   0.675139119  0.2006812061   0   0
#>   0.0833143570   0.390178011  0.0993080001   0   0
#>   0.1316951313   0.030650374  0.0696571923   0   0
#>   0.2828729256   0.572854201  0.1954918967   0   0
#>   0.1939443999   0.643905482  0.2685552847   0   0
#>   0.0993080001   0.393429782  0.0420464265   0   0
#>   0.0964163853   0.158183609  0.0084790498   0   0
#>   0.8985657152   0.903670627  0.8762494111   0   0
#>   0.6990782097   0.914880037  0.6864535876   0   0
#>   0.6671950352   0.729473419  0.6827132604   0   0
#>   0.2822566071   0.725098920  0.3745865022   0   0
#>   0.1666861925   0.453757201  0.1261186512   0   0
#>  -0.0085563660   0.315102944 -0.1992726409   0   0
#>   0.0839478312   0.284340446 -0.0085563660   0   0
#>   0.1074176070   0.453757201  0.1261186512   0   0
#>  -0.1318622524   0.158183609  0.0084790498   0   0
#>   0.5117992890   0.860953370  0.5062514678   0   0
#>   0.3745865022   0.721223409  0.3779714684   0   0
#>   0.1772632966   0.589522647  0.1939443999   0   0
#>   0.8366431750   1.147990397  0.7655081495   0   0
#>   0.7293102566   0.984930971  0.8366431750   0   0
#>   0.1630329391   0.542714010  0.1378973442   0   0
#>   0.1261186512   0.502656149  0.0500215262   0   0
#>   0.0420464265   0.464935123  0.0251737846   0   0
#>   0.6275860912   0.848775313  0.5784890018   0   0
#>   0.8762494111   0.842565794  0.7774471310   0   0
#>   0.8645750168   4.854912149 16.0861202396   0   1
#>   0.1145637502   0.542714010  0.1378973442   0   0
#>   0.4995886475   0.983790793  0.4423249695   0   0
#>   0.3691650848   0.725098920  0.3745865022   0   0
#>  -0.1992726409   0.205979838 -0.9912216861   0   0
#>  -0.9912216861   0.132018371 -0.1318622524   0   0
#>   0.0400098921   0.502656149  0.0500215262   0   0
#>   0.2136528994   0.675139119  0.2006812061   0   0
#>   0.0434844384   0.560220413  0.0691157586   0   0
#>   0.7355837701   3.038980051  0.5844730291   0   0
#>   0.6023832325   1.307698857  0.7355837701   0   0
#>   0.0691157586   0.542714010  0.1378973442   0   0
#>  -0.1584044526   0.464935123  0.0251737846   0   0
#>   0.7655081495   0.914880037  0.6864535876   0   0
#>   0.3297595935   0.745149290  0.3691650848   0   0
#>   0.2237705841   0.760872190  0.2136528994   0   0
#>   0.0753042952   0.654681996  0.1145637502   0   0
#>   0.4423249695   3.038980051  0.5844730291   0   0
#>   0.3512980536   0.931507142  0.3076440119   0   0
#>   0.1915150115   0.794773998  0.2237705841   0   0
#>   0.0803715375   0.672835922  0.0688932939   0   0
#>   0.0688932939   0.671149674  0.0753042952   0   0
#>  -0.8152939448   0.781496417 -0.0013995019   0   0
#>   0.1039261871   0.718904187  0.0803715375   0   0
#>   0.1881936703   0.931507142  0.3076440119   0   0
#>   0.1028515413   0.907003541  0.0008349779   0   0
#>  -0.9415037512   0.905270635 -3.6874064011   0   1
#>   0.4163803731   3.038980051  0.5844730291   0   0
#>   0.1989500013   0.931507142  0.3076440119   0   0
#>   0.1629252651   0.923481977  0.1881936703   0   0
#>   0.8106702558   0.083064969  0.8790191999   0   0
#>   0.7908216857   0.175822781  0.7394610191   0   0
#>   0.7799771677  -0.124634166  0.6994738874   0   0
#>   0.7186159374   0.180998220  0.7092828171   0   0
#>   0.8790191999   0.070055944  0.8790789153   0   0
#>   0.6990829869  -0.104097662  0.7003162055   0   0
#>   0.7092828171   0.230405204  0.6441943579   0   0
#>   0.6907834943   0.101921384  0.6326402488   0   0
#>   0.7394610191   0.177099467  0.7186159374   0   0
#>   0.6298744206   0.125252744  0.6347825967   0   0
#>   0.8157079654   0.180383021  0.9152047021   0   0
#>   0.7975826335   0.187019548  0.7908216857   0   0
#>   0.6333375438   0.230405204  0.6441943579   0   0
#>   0.6326402488   0.125252744  0.6347825967   0   0
#>   0.8068706782   0.365552827  0.7396781691   0   0
#>   0.6356207146   0.319288508  0.5543338553   0   0
#>   0.6351453586   0.236286488  0.5512640301   0   0
#>   0.7178211664   0.365552827  0.7396781691   0   0
#>   1.3933701040   0.361036766  0.9256121408   0   0
#>   0.9256121408   0.396450944  0.8667816936   0   0
#>   0.5234804690   0.319288508  0.5543338553   0   0
#>   0.7127550651   0.470839082  0.6393738436   0   0
#>   0.4985929248  -0.052375788  0.5694164030   0   0
#>   0.5694164030  -0.329270369  0.6232824374   0   0
#>   0.8649913514   0.465792572  0.8864208363   0   0
#>   0.4762892124   0.425639467  0.3883174636   0   0
#>   0.4336527479   0.259432618  0.2781598534   0   0
#>   0.4423389564   0.151237684  0.4484397034   0   0
#>   0.4235936905   0.218501657  0.4336527479   0   0
#>   0.7936255362   0.539835411  0.7806601539   0   0
#>   0.6393738436   0.473164962  0.6347267475   0   0
#>   0.6008518955   0.473164962  0.6347267475   0   0
#>   0.2175801213   0.108966443  0.1956476846   0   0
#>   0.8274312983   0.640154764  0.8066988095   0   0
#>   0.7506816560   0.621730753  0.7279295028   0   0
#>   1.1261440159   0.235472581  2.5995425937   0   0
#>   1.0007800203   0.564841696  0.8777250307   0   0
#>   0.8485385127   0.549342661  0.8274312983   0   0
#>   0.5742934213   0.551865799  0.5155346025   0   0
#>   0.6626696617   0.622963971  0.5669809950   0   0
#>   0.8308492207   0.683449598  0.8307138587   0   0
#>   2.5995425937   0.251587796  1.3933701040   0   0
#>   0.8859764590   0.633939327  0.9254220666   0   0
#>   0.4806021854   0.513985667  0.3656780048   0   0
#>   0.3645544514   0.425639467  0.3883174636   0   0
#>   0.2781598534   0.311814502  0.3086319207   0   0
#>   0.9301647676   0.699559257  0.9051157291   0   0
#>   0.6748233408   0.718369186  0.5496875223   0   0
#>   0.6726503000   0.644081519  0.6626696617   0   0
#>   0.5155346025   0.622963971  0.5669809950   0   0
#>   0.9051157291   0.709528438  0.8881456193   0   0
#>   0.8307138587   0.709528438  0.8881456193   0   0
#>   0.1956476846   0.078175507  0.1326426055   0   0
#>   0.3117509514  -1.462876966  0.5713592555   0   0
#>   0.6827132604   0.717856298  0.6748233408   0   0
#>   0.9517630346   0.747251701  0.9722482031   0   0
#>   0.8951211138   0.805892747  0.8600249653   0   0
#>   0.4513639159   0.721223409  0.3779714684   0   0
#>   0.2155034745   0.215471024  0.2225905033   0   0
#>   0.7165212542  -6.339093944  0.7755534958   0   1
#>   0.3583716585   0.513985667  0.3656780048   0   0
#>   0.2572280646   0.378855452  0.1306779067   0   0
#>   0.3558483032   0.571456866  0.2896483701   0   0
#>   2.7161947075   0.692974996  1.0625377188   0   0
#>   0.9722482031   0.697401483  0.9301647676   0   0
#>   0.2163138916   0.571456866  0.2896483701   0   0
#>   0.0634416660   0.319447620  0.0833143570   0   0
#>   0.7888192220   0.805892747  0.8600249653   0   0
#>   0.2685552847   0.640024691  0.2822566071   0   0
#>   0.0833143570   0.210216798  0.1585272842   0   0
#>   0.1585272842   0.210457757  0.0964163853   0   0
#>   0.1316951313   0.078175507  0.1326426055   0   0
#>   0.2828729256   0.640024691  0.2822566071   0   0
#>   0.1939443999   0.572854201  0.1954918967   0   0
#>   0.0993080001   0.393451912  0.1074176070   0   0
#>   0.0964163853   0.225907045  0.0839478312   0   0
#>   0.8985657152   0.853599085  0.9366277366   0   0
#>   0.7774471310   0.857453263  0.7293102566   0   0
#>   0.6990782097   0.857453263  0.7293102566   0   0
#>   0.6671950352   0.848775313  0.5784890018   0   0
#>   0.1666861925   0.523560320  0.1630329391   0   0
#>   0.0839478312   0.297686496  0.0634416660   0   0
#>  -0.1318622524   0.030650374  0.0696571923   0   0
#>   0.0696571923  -8.508252109 -0.1586824190   0   1
#>   0.5117992890   0.848775313  0.5784890018   0   0
#>   0.5062514678   0.745149290  0.3691650848   0   0
#>   0.1772632966   0.654681996  0.1145637502   0   0
#>   0.8366431750   0.942290874  0.8645750168   0   0
#>   0.1954918967   0.523560320  0.1630329391   0   0
#>   0.6275860912   0.954464591  0.6023832325   0   0
#>   0.8762494111   0.942290874  0.8645750168   0   0
#>   0.2006812061   0.654086168  0.1772632966   0   0
#>   0.6864535876   0.910507171  0.6275860912   0   0
#>   0.4995886475   0.954464591  0.6023832325   0   0
#>   0.0500215262   0.464935123  0.0251737846   0   0
#>  -0.1992726409   0.393429782  0.0420464265   0   0
#>  -0.9912216861   0.155147750 -3.2051055535   0   1
#>   0.0400098921   0.547836982 -0.1584044526   0   0
#>   0.2136528994   0.671149674  0.0753042952   0   0
#>   0.0434844384   0.703882571 -0.8152939448   0   0
#>   0.7355837701   1.147990397  0.7655081495   0   0
#>   0.0691157586   0.529833057  0.0400098921   0   0
#>  -0.1584044526   0.705255993 -0.9415037512   0   0
#>   0.3297595935   0.794773998  0.2237705841   0   0
#>   0.3512980536   0.983790793  0.4423249695   0   0
#>   0.1915150115   0.797432559  0.1039261871   0   0
#>   0.0803715375   0.781496417 -0.0013995019   0   0
#>   0.0688932939   0.661580007  0.0434844384   0   0
#>   0.1039261871   0.797912488  0.1028515413   0   0
#>   0.1881936703   0.826125270  0.1915150115   0   0
#>  -0.0013995019   0.797912488  0.1028515413   0   0
#>  -0.9415037512   0.703882571 -0.8152939448   0   0
#>   0.4163803731   1.657360372  0.1989500013   0   0
#>   0.1989500013   0.971539525  0.1629252651   0   0
#>   0.1629252651   0.907003541  0.0008349779   0   0
#>   0.0008349779   1.460886068 -5.7018164151   0   1
#>   0.8106702558  -0.002800944  0.7799771677   0   0
#>   0.8157079654   0.189619366  0.7975826335   0   0
#>   0.6356207146   0.370290518  0.7178211664   0   0
#>   0.5234804690   0.258679622  0.5100866163   0   0
#>   0.8649913514   0.456984927  0.8028858834   0   0
#>   0.4235936905   0.344578328  0.3645544514   0   0
#>   0.2175801213   0.198701472  0.2155034745   0   0
#>   0.7506816560   0.640154764  0.8066988095   0   0
#>   1.1261440159   0.517303962  1.0007800203   0   0
#>   0.8777250307   0.548382548  0.8485385127   0   0
#>   0.9517630346   0.727385127  0.8951211138   0   0
#>   0.4513639159   0.704133368  0.5340690850   0   0
#>   0.7165212542   0.077172183  0.1316951313   0   0
#>   0.2572280646   0.483194977  0.2163138916   0   0
#>   0.1306779067   0.348618466  0.1705593471   0   0
#>   2.7161947075   0.068871887  4.0992625687   0   1
#>   0.7783177150   0.752264414  0.7888192220   0   0
#>   0.6990782097   0.798178809  0.6671950352   0   0
#>   0.5062514678   0.886995564  0.4995886475   0   0
#>   0.3297595935   0.920047323  0.3512980536   0   0
#>   0.4163803731  10.439071804  0.4051823036   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.070143781 0.83964222
#>   [2,] 0.542551234 0.24072417
#>   [3,] 0.599582736 0.79189514
#>   [4,] 0.572632413 0.02372969
#>   [5,] 0.518065435 0.27019411
#>   [6,] 0.673577916 0.93020614
#>   [7,] 0.270893179 0.37547542
#>   [8,] 0.390131261 0.30561636
#>   [9,] 0.466931547 0.05231843
#>  [10,] 0.990561435 0.04325996
#>  [11,] 0.234709979 0.03029688
#>  [12,] 0.816843211 0.02638720
#>  [13,] 0.351100769 0.43875840
#>  [14,] 0.003008577 0.15346237
#>  [15,] 0.201543938 0.35722037
#>  [16,] 0.766635931 0.90105883
#>  [17,] 0.410980602 0.20116389
#>  [18,] 0.651974202 0.76515920
#>  [19,] 0.595711691 0.59374733
#>  [20,] 0.091417905 0.77806408
#>  [21,] 0.592714852 0.10314611
#>  [22,] 0.866091855 0.39847689
#>  [23,] 1.000000000 0.96129331
#>  [24,] 0.903812483 0.74080726
#>  [25,] 0.610342905 0.23497386
#>  [26,] 0.319237660 0.04393118
#>  [27,] 0.753835840 0.49330545
#>  [28,] 0.480508405 0.76738333
#>  [29,] 0.883498795 0.64071488
#>  [30,] 0.032388486 0.86739888
#>  [31,] 0.882338803 0.29034642
#>  [32,] 0.938375320 0.97713363
#>  [33,] 0.213829828 0.57049341
#>  [34,] 0.000000000 0.12739310
#>  [35,] 0.816900638 0.75454600
#>  [36,] 0.104036089 0.67461934
#>  [37,] 0.467790799 0.84408077
#>  [38,] 0.720867015 0.93427972
#>  [39,] 0.573821253 0.73025251
#>  [40,] 0.658950399 0.57979117
#>  [41,] 0.140786074 0.08612974
#>  [42,] 0.053190139 0.41324929
#>  [43,] 0.611324414 0.33014148
#>  [44,] 0.454586314 0.00000000
#>  [45,] 0.982831856 0.31672730
#>  [46,] 0.749136154 0.86191306
#>  [47,] 0.823380055 0.80599655
#>  [48,] 0.690093341 0.25756254
#>  [49,] 0.323451188 0.85017619
#>  [50,] 0.349026121 0.59885808
#>  [51,] 0.507162427 0.89565174
#>  [52,] 0.132884553 0.60421494
#>  [53,] 0.875472645 0.08908238
#>  [54,] 0.283732973 0.47694649
#>  [55,] 0.156119144 0.61600862
#>  [56,] 0.487981730 0.83209176
#>  [57,] 0.622209545 0.22632443
#>  [58,] 0.265850059 0.55100755
#>  [59,] 0.777539212 0.58027723
#>  [60,] 0.725735753 0.13813381
#>  [61,] 0.876678606 0.77303376
#>  [62,] 0.105457463 0.70847935
#>  [63,] 0.368763747 0.53386601
#>  [64,] 0.977919809 0.28392489
#>  [65,] 0.302313722 0.11877572
#>  [66,] 0.576280726 0.71259942
#>  [67,] 0.647820214 0.96154275
#>  [68,] 0.070505823 0.91851314
#>  [69,] 0.706251417 0.15076989
#>  [70,] 0.595488716 0.92977333
#>  [71,] 0.965549671 0.56659466
#>  [72,] 0.445291637 0.92352051
#>  [73,] 0.705394865 0.95260736
#>  [74,] 0.286456860 0.88635535
#>  [75,] 0.484345876 0.45653669
#>  [76,] 0.490663101 0.01629187
#>  [77,] 0.603334583 0.39959835
#>  [78,] 0.280207157 0.08667062
#>  [79,] 0.349692642 0.24001811
#>  [80,] 0.602433823 0.14978332
#>  [81,] 0.270574255 0.05295177
#>  [82,] 0.099359264 0.84091409
#>  [83,] 0.752038480 0.03256705
#>  [84,] 0.784232714 1.00000000
#>  [85,] 0.094660769 0.05640766
#>  [86,] 0.422863708 0.19442406
#>  [87,] 0.099657885 0.74995152
#>  [88,] 0.080814051 0.30616516
#>  [89,] 0.756399338 0.03590510
#>  [90,] 0.610027083 0.83764329
#>  [91,] 0.917220044 0.59831572
#>  [92,] 0.220433434 0.52630905
#>  [93,] 0.515410900 0.08347058
#>  [94,] 0.249443024 0.71729795
#>  [95,] 0.939158912 0.63658027
#>  [96,] 0.003129441 0.01036572
#>  [97,] 0.127190263 0.66622893
#>  [98,] 0.639866695 0.53272575
#>  [99,] 0.663711129 0.90070319
#> [100,] 0.154874862 0.64551367
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,20,82), [15,8,4]
#> 2: (94,20,87), [3,12,15]
#> 3: (20,30,87), [9,2,4]
#> 4: (20,1,30), [5,3,1]
#> 5: (1,68,30), [0,4,8]
#> 6: (94,62,97), [11,10,12]
#> 7: (87,36,62), [11,12,9]
#> 8: (1,82,68), [19,5,1]
#> 9: (87,30,36), [40,7,3]
#> 10: (94,97,100), [18,22,6]
#> 11: (97,62,36), [7,17,6]
#> 12: (94,87,62), [7,6,2]
#> 13: (52,55,100), [16,18,23]
#> 14: (82,94,74), [20,19,15]
#> 15: (94,82,20), [1,2,14]
#> 16: (100,55,33), [23,22,13]
#> 17: (97,36,52), [39,18,11]
#> 18: (52,100,97), [10,17,13]
#> 19: (82,74,68), [28,8,14]
#> 20: (49,74,94), [14,36,29]
#> 21: (58,33,92), [25,33,26]
#> 22: (100,33,94), [26,10,16]
#> 23: (33,55,52), [13,25,16]
#> 24: (50,94,58), [26,35,27]
#> 25: (92,33,52), [23,38,21]
#> 26: (58,94,33), [22,21,24]
#> 27: (94,50,28), [51,36,24]
#> 28: (72,68,74), [19,29,68]
#> 29: (72,74,49), [20,30,28]
#> 30: (49,37,72), [43,29,31]
#> 31: (28,37,49), [30,36,42]
#> 32: (58,54,63), [46,35,33]
#> 33: (54,58,92), [21,48,32]
#> 34: (66,39,28), [44,51,63]
#> 35: (58,63,50), [52,24,32]
#> 36: (94,28,49), [31,20,27]
#> 37: (15,92,42), [38,54,48]
#> 38: (42,92,52), [25,39,37]
#> 39: (42,52,36), [17,40,38]
#> 40: (42,36,30), [9,89,39]
#> 41: (37,56,51), [60,43,42]
#> 42: (56,37,28), [31,50,41]
#> 43: (37,51,72), [58,30,41]
#> 44: (3,28,39), [34,56,50]
#> 45: (75,63,13), [46,62,61]
#> 46: (13,63,54), [32,49,45]
#> 47: (7,54,15), [48,73,49]
#> 48: (15,54,92), [33,37,47]
#> 49: (7,13,54), [46,47,72]
#> 50: (3,56,28), [42,44,55]
#> 51: (66,28,50), [27,65,34]
#> 52: (50,63,19), [61,65,35]
#> 53: (41,15,88), [54,87,95]
#> 54: (88,15,42), [37,88,53]
#> 55: (90,56,3), [50,67,60]
#> 56: (18,3,39), [44,63,67]
#> 57: (72,70,67), [70,68,58]
#> 58: (70,72,51), [43,59,57]
#> 59: (51,90,70), [69,58,60]
#> 60: (90,51,56), [41,55,59]
#> 61: (19,63,75), [45,78,52]
#> 62: (13,8,75), [99,45,72]
#> 63: (18,39,66), [34,77,56]
#> 64: (40,66,19), [65,79,77]
#> 65: (50,19,66), [64,51,52]
#> 66: (99,90,18), [67,84,69]
#> 67: (3,18,90), [66,55,56]
#> 68: (72,67,68), [105,28,57]
#> 69: (70,90,99), [66,70,59]
#> 70: (70,99,67), [74,57,69]
#> 71: (77,98,75), [78,86,94]
#> 72: (13,7,8), [81,62,49]
#> 73: (7,15,79), [109,81,47]
#> 74: (6,67,99), [70,83,82]
#> 75: (38,73,6), [82,83,107]
#> 76: (59,18,40), [77,116,91]
#> 77: (40,18,66), [63,64,76]
#> 78: (19,75,98), [71,79,61]
#> 79: (19,98,40), [111,64,78]
#> 80: (17,8,79), [81,108,97]
#> 81: (7,79,8), [80,72,73]
#> 82: (73,67,6), [74,75,106]
#> 83: (38,6,99), [74,90,75]
#> 84: (99,18,46), [110,90,66]
#> 85: (27,77,43), [103,139,94]
#> 86: (75,5,77), [103,71,99]
#> 87: (41,88,14), [88,98,53]
#> 88: (14,88,42), [54,89,87]
#> 89: (14,42,30), [40,96,88]
#> 90: (99,46,38), [93,83,84]
#> 91: (59,35,18), [110,76,129]
#> 92: (38,16,84), [114,107,93]
#> 93: (16,38,46), [90,102,92]
#> 94: (27,98,77), [71,85,111]
#> 95: (15,41,65), [119,109,53]
#> 96: (34,14,30), [89,0,121]
#> 97: (5,8,17), [80,100,99]
#> 98: (14,85,41), [120,87,121]
#> 99: (75,8,5), [97,86,62]
#> 100: (86,5,17), [97,101,112]
#> 101: (17,65,86), [134,100,108]
#> 102: (46,47,16), [126,93,115]
#> 103: (43,77,5), [86,104,85]
#> 104: (5,2,43), [122,103,112]
#> 105: (84,68,67), [68,106,0]
#> 106: (84,67,73), [82,107,105]
#> 107: (38,84,73), [106,75,92]
#> 108: (65,17,79), [80,109,101]
#> 109: (15,65,79), [108,73,95]
#> 110: (35,46,18), [84,91,115]
#> 111: (98,27,40), [116,79,94]
#> 112: (5,86,2), [131,104,100]
#> 113: (78,81,26), [132,118,133]
#> 114: (16,32,84), [0,92,126]
#> 115: (46,35,47), [127,102,110]
#> 116: (40,27,59), [137,76,111]
#> 117: (25,57,48), [152,130,123]
#> 118: (78,26,65), [124,119,113]
#> 119: (78,65,41), [95,125,118]
#> 120: (11,41,85), [98,135,125]
#> 121: (14,34,85), [136,98,96]
#> 122: (43,2,25), [143,130,104]
#> 123: (80,57,25), [117,143,140]
#> 124: (65,26,9), [146,134,118]
#> 125: (78,41,11), [120,133,119]
#> 126: (16,47,32), [149,114,102]
#> 127: (61,47,35), [115,142,149]
#> 128: (35,29,24), [155,142,129]
#> 129: (29,35,59), [91,148,128]
#> 130: (25,48,43), [139,122,117]
#> 131: (2,86,93), [145,144,112]
#> 132: (81,11,26), [159,113,133]
#> 133: (78,11,81), [132,113,125]
#> 134: (65,9,86), [145,101,124]
#> 135: (96,11,85), [120,136,160]
#> 136: (96,85,34), [121,0,135]
#> 137: (59,27,91), [138,148,116]
#> 138: (22,91,27), [137,157,156]
#> 139: (48,27,43), [85,130,157]
#> 140: (69,57,80), [123,151,152]
#> 141: (61,24,23), [168,150,142]
#> 142: (35,24,61), [141,127,128]
#> 143: (80,25,2), [122,144,123]
#> 144: (2,93,80), [153,143,131]
#> 145: (86,9,93), [158,131,134]
#> 146: (26,44,9), [154,124,159]
#> 147: (95,29,91), [148,165,155]
#> 148: (59,91,29), [147,129,137]
#> 149: (32,47,61), [127,150,126]
#> 150: (61,23,32), [0,149,141]
#> 151: (69,80,21), [153,171,140]
#> 152: (69,48,57), [117,140,162]
#> 153: (21,80,93), [144,166,151]
#> 154: (76,9,44), [146,167,158]
#> 155: (95,24,29), [128,147,168]
#> 156: (91,22,71), [172,165,138]
#> 157: (22,27,48), [139,169,138]
#> 158: (76,93,9), [145,154,161]
#> 159: (26,11,44), [160,146,132]
#> 160: (44,11,96), [135,0,159]
#> 161: (4,93,76), [158,167,166]
#> 162: (60,48,69), [152,171,170]
#> 163: (83,21,4), [166,178,177]
#> 164: (95,71,23), [173,168,165]
#> 165: (91,71,95), [164,147,156]
#> 166: (4,21,93), [153,161,163]
#> 167: (4,76,44), [154,184,161]
#> 168: (23,24,95), [155,164,141]
#> 169: (31,22,48), [157,170,174]
#> 170: (31,48,60), [162,175,169]
#> 171: (60,69,21), [151,177,162]
#> 172: (22,45,71), [173,156,174]
#> 173: (45,23,71), [164,172,185]
#> 174: (22,31,45), [180,172,169]
#> 175: (53,31,60), [170,179,181]
#> 176: (89,60,83), [177,183,179]
#> 177: (83,60,21), [171,163,176]
#> 178: (4,12,83), [183,163,184]
#> 179: (53,60,89), [176,182,175]
#> 180: (64,45,31), [174,181,186]
#> 181: (53,64,31), [180,175,187]
#> 182: (89,12,53), [188,179,183]
#> 183: (89,83,12), [178,182,176]
#> 184: (4,44,12), [0,178,167]
#> 185: (10,23,45), [173,186,0]
#> 186: (10,45,64), [180,187,185]
#> 187: (10,64,53), [181,188,186]
#> 188: (10,53,12), [182,0,187]
#> boundary nodes:  68 30 34 96 44 12 10 23 32 84 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.063349734 0.029295278 0.135347902 0.046860371 0.063762053 0.047512234 
#>                                                                         
#> 0.033889832 0.082789699 0.205663149 0.034576855 0.041875721 0.029531281 
#>                                                                         
#> 0.173061922 0.073499370 0.076085559 0.218332257 0.044675110 0.151062478 
#>                                                                         
#> 0.026056419 0.167097833 0.087688537 0.213748337 0.051744678 0.082564434 
#>                                                                         
#> 0.144468257 0.102323328 0.051698028 0.100429007 0.067923071 0.177543310 
#>                                                                         
#> 0.201873297 0.117192728 0.066391035 0.077744672 0.067857670 0.096733845 
#>                                                                         
#> 0.145221838 0.080271374 0.077438686 0.065138559 0.234715024 0.131081424 
#>                                                                         
#> 0.158581492 0.118619363 0.066809184 0.061219949 0.093270078 0.139064916 
#>                                                                         
#> 0.177913861 0.017823605 0.120429980 0.083746667 0.578915757 0.082816228 
#>                                                                         
#> 0.062880629 0.173180536 0.138195547 0.188903169 0.038911859 0.185499256 
#>                                                                         
#> 0.092152229 0.050787215 0.077061090 0.040564055 0.031109123 0.137119488 
#>                                                                         
#> 0.069914910 0.154971884 0.171382127 0.110589741 0.454625734 0.077692198 
#>                                                                         
#> 0.165267569 0.100493789 0.138048907 0.164793632 0.714540770 0.106512974 
#>                                                                         
#> 0.054872081 0.132747462 0.127406175 0.142319960 0.110710131 0.112806052 
#>                                                                         
#> 0.058263898 0.091986243 0.072566279 0.130170966 0.128386243 0.128330838 
#>                                                                         
#> 0.049492402 0.155829439 0.051856932 0.090143997 0.074726362 0.076734111 
#>                                                                         
#> 0.118319878 0.068035114 0.014684337 0.147931843 0.109265764 0.206183923 
#>                                                                         
#> 0.102132903 0.178658694 0.107108571 0.144484107 0.085620225 0.042420396 
#>                                                                         
#> 0.148781670 0.142462465 0.194220378 0.216637918 0.079054546 0.240553820 
#>                                                                         
#> 0.042128312 0.109409177 0.057625554 0.053755209 0.054173901 0.062668024 
#>                                                                         
#> 0.063627955 0.047639138 0.195334184 0.244208568 0.221953833 0.232436652 
#>                                                                         
#> 0.071592128 0.108008120 0.081916255 0.396199258 0.074796105 0.079768586 
#>                                                                         
#> 0.039584840 0.110057418 0.225460262 0.124633875 0.123120168 0.250464313 
#>                                                                         
#> 0.103898050 0.218387711 0.108794185 0.137545263 0.065098723 0.106728463 
#>                                                                         
#> 0.095796476 0.085837613 0.363216662 0.644794613 0.033168150 0.220134109 
#>                                                                         
#> 0.029243155 0.153303110 0.107075575 0.132490686 0.078871753 0.075459302 
#>                                                                         
#> 0.118727077 0.024627519 0.144255361 0.046788447 0.192536797 0.169305821 
#>                                                                         
#> 0.095413460 0.062274870 0.152098114 0.095957960 0.055550005 0.236431292 
#>                                                                         
#> 0.163124909 0.142212815 0.104331490 0.110334106 0.206927719 0.266269811 
#>                                                                         
#> 0.064882127 0.134425913 0.071711674 0.170140513 0.102280196 0.121571070 
#>                                                                         
#> 0.254129391 0.246738506 0.228061904 0.046925238 0.085594159 0.206066774 
#>                                                                         
#> 0.094687993 0.122171550 0.176717582 0.064760394 0.136051198 0.374819263 
#>                                                                         
#> 0.074157741 0.131910149 0.138745032 0.156710045 0.047464227 0.118589810 
#>                                                                         
#> 0.156436264 0.075321294 0.066288600 0.093819557 0.153350327 0.264589174 
#>                                                                         
#> 0.223490306 0.091292756 0.042879292 0.177289730 0.258863673 0.740714224 
#>                                                                         
#> 0.189368866 0.013661387 0.110864439 0.141729778 0.023985199 0.038314903 
#>                                                                         
#> 0.057879527 0.092945215 0.082887758 0.038979962 0.139422132 0.133490602 
#>                                                                         
#> 0.095172679 0.085556894 0.177526335 0.072443021 0.187834481 0.062571578 
#>                                                                         
#> 0.087991033 0.121984191 0.159578472 0.035067847 0.102458991 0.117069219 
#>                                                                         
#> 0.140840768 0.146948525 0.103822282 0.225055164 0.108886009 0.044107705 
#>                                                                         
#> 0.213221109 0.113011173 0.055813473 0.057809846 0.043140462 0.142299708 
#>                                                                         
#> 0.451575859 0.082306071 0.023223044 0.179623595 0.330363740 0.082724096 
#>                                                                         
#> 0.120407556 0.195020769 0.142517448 0.157566430 0.005491782 0.174256771 
#>                                                                         
#> 0.130408111 0.201381123 0.061188660 0.244225257 0.273576551 0.240996717 
#>                                                                         
#> 0.123875434 0.174535702 0.065149488 0.194437870 0.154741059 0.076189496 
#>                                                                         
#> 0.023482118 0.102166498 0.277815854 0.058819011 0.150327177 0.118093680 
#>                                                                         
#> 0.056554584 0.120594057 0.026242302 0.121673619 0.136368256 0.718363546 
#>                                                             
#> 0.126963419 0.131881915 0.206275667 0.109344235 0.918081868 
#> 
#> $alpha
#> [1] 0.2999167
#> 
#> attr(,"class")
#> [1] "scree"