Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1         my1
#>   100   27 0.25279034 0.61524234 0.27095963 0.58710189  0.254246989  0.59624699
#>    58   18 0.23977129 0.49304972 0.31630352 0.50580654  0.277506374  0.50261398
#>     1   27 0.27313740 0.59377765 0.27095963 0.58710189  0.335424824  0.56976513
#>    18   27 0.31630352 0.50580654 0.27095963 0.58710189  0.281445142  0.53965702
#>   100    1 0.25279034 0.61524234 0.27313740 0.59377765  0.328360161  0.66650121
#>     1   18 0.27313740 0.59377765 0.31630352 0.50580654  0.338939731  0.57148984
#>   100   62 0.25279034 0.61524234 0.25668405 0.75308446  0.227525977  0.68493205
#>   100   32 0.25279034 0.61524234 0.39822756 0.60770651  0.329166389  0.68206096
#>    58   54 0.23977129 0.49304972 0.24024803 0.51389348  0.172623531  0.50501286
#>    77   54 0.10884141 0.52917438 0.24024803 0.51389348  0.180499110  0.57273815
#>   100   98 0.25279034 0.61524234 0.21219355 0.75745695  0.164705036  0.66699910
#>    58   77 0.23977129 0.49304972 0.10884141 0.52917438  0.169052346  0.49206949
#>    43   18 0.25156404 0.39083295 0.31630352 0.50580654  0.286828092  0.44669001
#>    28   43 0.23250161 0.39311987 0.25156404 0.39083295  0.248060213  0.44221737
#>    58   31 0.23977129 0.49304972 0.10009925 0.50780485  0.164428631  0.44830138
#>     2   32 0.53215701 0.56300956 0.39822756 0.60770651  0.445599076  0.52664916
#>    77   72 0.10884141 0.52917438 0.06969028 0.70454877  0.081143758  0.61504838
#>    77  100 0.10884141 0.52917438 0.25279034 0.61524234  0.146480513  0.62963436
#>    28   31 0.23250161 0.39311987 0.10009925 0.50780485  0.163742335  0.44750906
#>    77   37 0.10884141 0.52917438 0.01541101 0.67686026  0.019480517  0.57603841
#>    31   77 0.10009925 0.50780485 0.10884141 0.52917438  0.041044479  0.54443676
#>    20   31 0.07613862 0.44410148 0.10009925 0.50780485  0.069357571  0.48300984
#>    37   72 0.01541101 0.67686026 0.06969028 0.70454877  0.044967799  0.68596604
#>    79   26 0.05969249 0.79183304 0.01650653 0.69806872  0.037761987  0.74510634
#>    98   79 0.21219355 0.75745695 0.05969249 0.79183304  0.131710271  0.75586744
#>    62   11 0.25668405 0.75308446 0.46772692 0.77597324  0.358248304  0.80101551
#>    98   62 0.21219355 0.75745695 0.25668405 0.75308446  0.240755365  0.81954240
#>    62   32 0.25668405 0.75308446 0.39822756 0.60770651  0.367139787  0.71903278
#>     2   19 0.53215701 0.56300956 0.53935681 0.66877789  0.476724208  0.61991216
#>    43    9 0.25156404 0.39083295 0.49390466 0.44614497  0.378017170  0.39534314
#>    98   61 0.21219355 0.75745695 0.27636151 0.87787407  0.192003729  0.84552121
#>    32   18 0.39822756 0.60770651 0.31630352 0.50580654  0.416297628  0.50929678
#>    19   11 0.53935681 0.66877789 0.46772692 0.77597324  0.448345463  0.68549232
#>    41   91 0.58101720 0.52166327 0.56016141 0.61328388  0.579592912  0.56952308
#>     9   41 0.49390466 0.44614497 0.58101720 0.52166327  0.522117759  0.50160292
#>    90   61 0.10350959 0.86373467 0.27636151 0.87787407  0.188643467  0.88659991
#>    85   48 0.50360695 0.89597181 0.38983595 0.96886240  0.425512132  0.89931262
#>    11   15 0.46772692 0.77597324 0.57502159 0.70988110  0.544240644  0.78004870
#>    19   91 0.53935681 0.66877789 0.56016141 0.61328388  0.593816893  0.65754806
#>    91   42 0.56016141 0.61328388 0.62572718 0.54585925  0.659185558  0.64398660
#>    23   34 0.51295881 0.36245916 0.53623014 0.40101340  0.491652447  0.40162008
#>    43   23 0.25156404 0.39083295 0.51295881 0.36245916  0.382268895  0.37671490
#>    61   11 0.27636151 0.87787407 0.46772692 0.77597324  0.383594318  0.84861425
#>    79   90 0.05969249 0.79183304 0.10350959 0.86373467  0.060444630  0.84067665
#>    90   60 0.10350959 0.86373467 0.24482205 0.96788267  0.169839424  0.92167891
#>    70   92 0.30623926 0.23028324 0.38096979 0.21219897  0.372970926  0.34259340
#>    28   76 0.23250161 0.39311987 0.03483506 0.25021822  0.151036395  0.29764493
#>    28   20 0.23250161 0.39311987 0.07613862 0.44410148  0.126117877  0.33211310
#>     3   60 0.26211580 0.93530732 0.24482205 0.96788267  0.177510827  0.91127003
#>    48   61 0.38983595 0.96886240 0.27636151 0.87787407  0.384399556  0.85938935
#>    41   42 0.58101720 0.52166327 0.62572718 0.54585925  0.583533323  0.57042005
#>    57   41 0.63685096 0.52148925 0.58101720 0.52166327  0.608822571  0.48579795
#>    42   57 0.62572718 0.54585925 0.63685096 0.52148925  0.682242010  0.55693191
#>    23   94 0.51295881 0.36245916 0.61786834 0.29904727  0.577080757  0.35005557
#>     3   61 0.26211580 0.93530732 0.27636151 0.87787407  0.333725897  0.92258607
#>    86   85 0.58004779 0.84778943 0.50360695 0.89597181  0.539074003  0.86751243
#>    63   90 0.03866075 0.88439994 0.10350959 0.86373467  0.081726523  0.90746051
#>    39    5 0.29821298 0.99007223 0.08853174 0.96643892  0.164181997  1.23724046
#>    39    3 0.29821298 0.99007223 0.26211580 0.93530732  0.332200797  0.92839104
#>    92   23 0.38096979 0.21219897 0.51295881 0.36245916  0.379042518  0.34699177
#>    63   25 0.03866075 0.88439994 0.08865039 0.95581866  0.006364631  0.96021017
#>     5   25 0.08853174 0.96643892 0.08865039 0.95581866  0.166724556  0.96200169
#>    86   15 0.58004779 0.84778943 0.57502159 0.70988110  0.586491216  0.77850884
#>    41   34 0.58101720 0.52166327 0.53623014 0.40101340  0.596722385  0.44719551
#>    30   16 0.38262948 0.13872288 0.46436137 0.17320660  0.414953766  0.17620981
#>    70   28 0.30623926 0.23028324 0.23250161 0.39311987  0.180621007  0.27151297
#>    55   76 0.03053718 0.47574207 0.03483506 0.25021822 -0.014549717  0.36207996
#>    12   76 0.01832299 0.21990683 0.03483506 0.25021822 -0.031864025  0.26689924
#>    10   86 0.55680588 0.94135654 0.58004779 0.84778943  0.618306894  0.90696311
#>    36   11 0.49398500 0.86161078 0.46772692 0.77597324  0.528884673  0.80406551
#>    99   57 0.72844298 0.52255169 0.63685096 0.52148925  0.682477048  0.53666964
#>    70   40 0.30623926 0.23028324 0.25222679 0.15918632  0.303342199  0.17641899
#>    74   40 0.06634976 0.20501546 0.25222679 0.15918632  0.180834243  0.26948853
#>    49   48 0.55965052 0.99180720 0.38983595 0.96886240  0.475209655  0.97688281
#>    65    7 0.70110926 0.75098393 0.62676203 0.83525018  0.627974169  0.76138862
#>    74   76 0.06634976 0.20501546 0.03483506 0.25021822  0.046054682  0.22445321
#>    37   97 0.01541101 0.67686026 0.00000000 0.20597545 -1.231290767  0.48196744
#>    10   49 0.55680588 0.94135654 0.55965052 0.99180720  0.485845385  0.97066314
#>     8   99 0.70047794 0.48551071 0.72844298 0.52255169  0.685823263  0.52565160
#>    34   94 0.53623014 0.40101340 0.61786834 0.29904727  0.637863179  0.39872043
#>    13   40 0.35689790 0.16997405 0.25222679 0.15918632  0.303768693  0.17228081
#>    74   97 0.06634976 0.20501546 0.00000000 0.20597545  0.032477218  0.15727638
#>    12   97 0.01832299 0.21990683 0.00000000 0.20597545  0.032832167  0.18180871
#>    94    8 0.61786834 0.29904727 0.70047794 0.48551071  0.648890014  0.39683476
#>    13   30 0.35689790 0.16997405 0.38262948 0.13872288  0.395788679  0.17577691
#>    15   91 0.57502159 0.70988110 0.56016141 0.61328388  0.656345118  0.64792894
#>    17   30 0.26658097 0.13426625 0.38262948 0.13872288  0.325324753  0.11775835
#>    42   99 0.62572718 0.54585925 0.72844298 0.52255169  0.697515482  0.62424171
#>     6   94 0.81485728 0.40574580 0.61786834 0.29904727  0.715821005  0.35339683
#>    17   40 0.26658097 0.13426625 0.25222679 0.15918632  0.175853333  0.09860044
#>     7   52 0.62676203 0.83525018 0.68382937 0.93845658  0.621968605  0.90528139
#>    33   65 0.71965599 0.72819923 0.70110926 0.75098393  0.650871200  0.69114934
#>    67   39 0.57652391 1.00000000 0.29821298 0.99007223  0.425386344  1.33093726
#>    86    7 0.58004779 0.84778943 0.62676203 0.83525018  0.620594941  0.90556025
#>    49   67 0.55965052 0.99180720 0.57652391 1.00000000  0.435837340  1.26827747
#>    67   10 0.57652391 1.00000000 0.55680588 0.94135654  0.604319652  0.95801740
#>    33   24 0.71965599 0.72819923 0.80172119 0.59931499  0.790574148  0.68278631
#>    33   99 0.71965599 0.72819923 0.72844298 0.52255169  0.698709356  0.62429272
#>     6    8 0.81485728 0.40574580 0.70047794 0.48551071  0.705996758  0.37153455
#>    16   73 0.46436137 0.17320660 0.45099809 0.06705591  0.511082506  0.11340840
#>    17   68 0.26658097 0.13426625 0.08940749 0.14366340  0.144823391 -0.48643640
#>     4   65 0.81287247 0.84993605 0.70110926 0.75098393  0.743870089  0.81527949
#>    94   16 0.61786834 0.29904727 0.46436137 0.17320660  0.569394134  0.20163040
#>    99   75 0.72844298 0.52255169 0.83667138 0.46273541  0.792433482  0.51051320
#>    59   24 0.86546581 0.72531659 0.80172119 0.59931499  0.907881529  0.62473323
#>    44   96 0.72156907 0.21886967 0.79477335 0.26702483  0.768742554  0.22687695
#>    44   71 0.72156907 0.21886967 0.72132648 0.23329331  0.643282977  0.22476682
#>    44   94 0.72156907 0.21886967 0.61786834 0.29904727  0.619124567  0.19352063
#>    24   99 0.80172119 0.59931499 0.72844298 0.52255169  0.800234097  0.52737725
#>    84   75 0.84304143 0.46954316 0.83667138 0.46273541  0.906620097  0.40366810
#>    52    4 0.68382937 0.93845658 0.81287247 0.84993605  0.761099349  0.91278067
#>    66   59 0.80333831 0.76602569 0.86546581 0.72531659  0.868688583  0.79799693
#>    67   21 0.57652391 1.00000000 0.65369049 0.96789982  0.657567890  1.08602245
#>    38   21 0.72624674 0.98983927 0.65369049 0.96789982  0.690887992  0.97582906
#>    66   24 0.80333831 0.76602569 0.80172119 0.59931499  0.793180158  0.68276103
#>    94   71 0.61786834 0.29904727 0.72132648 0.23329331  0.716676562  0.34024525
#>    29   94 0.82034856 0.31311546 0.61786834 0.29904727  0.716718624  0.34047762
#>    81   96 0.78903147 0.11570666 0.79477335 0.26702483  0.792107624  0.19135796
#>    46   84 0.95301725 0.52534072 0.84304143 0.46954316  0.920679451  0.45279905
#>    38   52 0.72624674 0.98983927 0.68382937 0.93845658  0.761767108  0.91731702
#>     4   69 0.81287247 0.84993605 0.92640324 0.75370431  0.871866576  0.80444954
#>     4   66 0.81287247 0.84993605 0.80333831 0.76602569  0.868995494  0.80106234
#>    46   45 0.95301725 0.52534072 0.96681225 0.52913808  0.930367604  0.63457786
#>    78   84 0.98622129 0.49765765 0.84304143 0.46954316  0.926194716  0.42471122
#>    45   69 0.96681225 0.52913808 0.92640324 0.75370431  0.942014467  0.64059467
#>    95   35 0.91863489 0.22502469 0.95877349 0.31166965  0.889339838  0.29121539
#>    22   81 0.65528720 0.00000000 0.78903147 0.11570666  0.687446077  0.09797806
#>    81   44 0.78903147 0.11570666 0.72156907 0.21886967  0.684957411  0.12128817
#>    64   73 0.58440755 0.09386031 0.45099809 0.06705591  0.537008051 -0.01562687
#>    73   68 0.45099809 0.06705591 0.08940749 0.14366340 -0.287804673 -2.52845908
#>    75    6 0.83667138 0.46273541 0.81485728 0.40574580  0.906380643  0.40338283
#>     4   82 0.81287247 0.84993605 0.94123267 0.83160966  0.879105483  0.85515171
#>    45   78 0.96681225 0.52913808 0.98622129 0.49765765  1.161283643  0.62731465
#>    81   83 0.78903147 0.11570666 0.90096643 0.21818239  0.823747759  0.19015735
#>    51   47 0.93378562 0.89288892 0.99575809 0.93181886  0.910998877  0.99795497
#>    14    4 0.84249048 0.91510142 0.81287247 0.84993605  0.877413249  0.85991538
#>    82   50 0.94123267 0.83160966 0.96894484 0.79552953  0.974026066  0.82811483
#>    82   69 0.94123267 0.83160966 0.92640324 0.75370431  0.929044410  0.79356564
#>    35   93 0.95877349 0.31166965 0.97691446 0.34501256  0.893430520  0.36882745
#>    88   87 0.85328041 0.07133669 0.81958807 0.10378424  0.767879075  0.01637529
#>    51   82 0.93378562 0.89288892 0.94123267 0.83160966  0.928793530  0.86119012
#>    96   83 0.79477335 0.26702483 0.90096643 0.21818239  0.857066932  0.26259979
#>    83   95 0.90096643 0.21818239 0.91863489 0.22502469  0.884803015  0.28615348
#>    53   95 0.96867742 0.19097218 0.91863489 0.22502469  0.925437649  0.18122507
#>    29   35 0.82034856 0.31311546 0.95877349 0.31166965  0.890097214  0.36372821
#>    56   50 0.96075306 0.85830503 0.96894484 0.79552953  1.019071156  0.83399291
#>    50   69 0.96894484 0.79552953 0.92640324 0.75370431  0.965051285  0.75694203
#>    88   53 0.85328041 0.07133669 0.96867742 0.19097218  0.906898148  0.13509063
#>    22   88 0.65528720 0.00000000 0.85328041 0.07133669  0.763512963  0.01005304
#>    47   38 0.99575809 0.93181886 0.72624674 0.98983927  0.896873786  1.12745564
#>    51   56 0.93378562 0.89288892 0.96075306 0.85830503  0.974515636  0.89684279
#>    56   80 0.96075306 0.85830503 1.00000000 0.77375696  1.063345298  0.85454482
#>    78   93 0.98622129 0.49765765 0.97691446 0.34501256 21.497736032 -0.82954350
#>    54  100 0.24024803 0.51389348 0.25279034 0.61524234  0.211831279  0.56886067
#>    54   18 0.24024803 0.51389348 0.31630352 0.50580654  0.277506374  0.50261398
#>    54   27 0.24024803 0.51389348 0.27095963 0.58710189  0.281445142  0.53965702
#>    32    1 0.39822756 0.60770651 0.27313740 0.59377765  0.328360161  0.66650121
#>    72   98 0.06969028 0.70454877 0.21219355 0.75745695  0.164705036  0.66699910
#>    58   43 0.23977129 0.49304972 0.25156404 0.39083295  0.248060213  0.44221737
#>     9   32 0.49390466 0.44614497 0.39822756 0.60770651  0.445599076  0.52664916
#>    72  100 0.06969028 0.70454877 0.25279034 0.61524234  0.146480513  0.62963436
#>    55   37 0.03053718 0.47574207 0.01541101 0.67686026  0.019480517  0.57603841
#>    55   77 0.03053718 0.47574207 0.10884141 0.52917438  0.041044479  0.54443676
#>    55   31 0.03053718 0.47574207 0.10009925 0.50780485  0.069357571  0.48300984
#>    26   72 0.01650653 0.69806872 0.06969028 0.70454877  0.044967799  0.68596604
#>    72   79 0.06969028 0.70454877 0.05969249 0.79183304  0.131710271  0.75586744
#>    61   62 0.27636151 0.87787407 0.25668405 0.75308446  0.240755365  0.81954240
#>    11   32 0.46772692 0.77597324 0.39822756 0.60770651  0.367139787  0.71903278
#>    32   19 0.39822756 0.60770651 0.53935681 0.66877789  0.476724208  0.61991216
#>    18    9 0.31630352 0.50580654 0.49390466 0.44614497  0.378017170  0.39534314
#>    91    2 0.56016141 0.61328388 0.53215701 0.56300956  0.490860446  0.61894988
#>    90   98 0.10350959 0.86373467 0.21219355 0.75745695  0.139911321  0.79224938
#>     2   41 0.53215701 0.56300956 0.58101720 0.52166327  0.522117759  0.50160292
#>    36   48 0.49398500 0.86161078 0.38983595 0.96886240  0.425512132  0.89931262
#>    15   19 0.57502159 0.70988110 0.53935681 0.66877789  0.512333484  0.72825028
#>    33   42 0.71965599 0.72819923 0.62572718 0.54585925  0.659185558  0.64398660
#>    34    9 0.53623014 0.40101340 0.49390466 0.44614497  0.556263948  0.46221430
#>     9   23 0.49390466 0.44614497 0.51295881 0.36245916  0.382268895  0.37671490
#>    63   79 0.03866075 0.88439994 0.05969249 0.79183304 -0.093698531  0.80565442
#>    25   60 0.08865039 0.95581866 0.24482205 0.96788267  0.169839424  0.92167891
#>    43   92 0.25156404 0.39083295 0.38096979 0.21219897  0.372970926  0.34259340
#>    76   20 0.03483506 0.25021822 0.07613862 0.44410148  0.126117877  0.33211310
#>    36   61 0.49398500 0.86161078 0.27636151 0.87787407  0.384399556  0.85938935
#>    89   41 0.65212562 0.49937387 0.58101720 0.52166327  0.608822571  0.48579795
#>    36   85 0.49398500 0.86161078 0.50360695 0.89597181  0.539074003  0.86751243
#>    25   90 0.08865039 0.95581866 0.10350959 0.86373467  0.081726523  0.90746051
#>    60    5 0.24482205 0.96788267 0.08853174 0.96643892  0.164181997  1.23724046
#>    48    3 0.38983595 0.96886240 0.26211580 0.93530732  0.332200797  0.92839104
#>     7   15 0.62676203 0.83525018 0.57502159 0.70988110  0.586491216  0.77850884
#>    89   34 0.65212562 0.49937387 0.53623014 0.40101340  0.596722385  0.44719551
#>    92   16 0.38096979 0.21219897 0.46436137 0.17320660  0.414953766  0.17620981
#>    74   28 0.06634976 0.20501546 0.23250161 0.39311987  0.180621007  0.27151297
#>    97   76 0.00000000 0.20597545 0.03483506 0.25021822 -0.031864025  0.26689924
#>    21   86 0.65369049 0.96789982 0.58004779 0.84778943  0.618306894  0.90696311
#>    86   11 0.58004779 0.84778943 0.46772692 0.77597324  0.528884673  0.80406551
#>    89   57 0.65212562 0.49937387 0.63685096 0.52148925  0.682477048  0.53666964
#>    13   70 0.35689790 0.16997405 0.30623926 0.23028324  0.340262711  0.20743156
#>    55   97 0.03053718 0.47574207 0.00000000 0.20597545 -1.231290767  0.48196744
#>    85   49 0.50360695 0.89597181 0.55965052 0.99180720  0.485845385  0.97066314
#>    89   99 0.65212562 0.49937387 0.72844298 0.52255169  0.685823263  0.52565160
#>    89   94 0.65212562 0.49937387 0.61786834 0.29904727  0.637863179  0.39872043
#>    68   74 0.08940749 0.14366340 0.06634976 0.20501546  0.165493215  0.20726732
#>    68   97 0.08940749 0.14366340 0.00000000 0.20597545  0.032477218  0.15727638
#>    89    8 0.65212562 0.49937387 0.70047794 0.48551071  0.648890014  0.39683476
#>    16   23 0.46436137 0.17320660 0.51295881 0.36245916  0.522179191  0.25922564
#>    92   30 0.38096979 0.21219897 0.38262948 0.13872288  0.395788679  0.17577691
#>    33   91 0.71965599 0.72819923 0.56016141 0.61328388  0.656345118  0.64792894
#>    68   40 0.08940749 0.14366340 0.25222679 0.15918632  0.175853333  0.09860044
#>    21   52 0.65369049 0.96789982 0.68382937 0.93845658  0.621968605  0.90528139
#>    15   65 0.57502159 0.70988110 0.70110926 0.75098393  0.650871200  0.69114934
#>    48   39 0.38983595 0.96886240 0.29821298 0.99007223  0.425386344  1.33093726
#>    21    7 0.65369049 0.96789982 0.62676203 0.83525018  0.620594941  0.90556025
#>    48   67 0.38983595 0.96886240 0.57652391 1.00000000  0.435837340  1.26827747
#>     4    7 0.81287247 0.84993605 0.62676203 0.83525018  0.719109033  0.85156819
#>    21   10 0.65369049 0.96789982 0.55680588 0.94135654  0.604319652  0.95801740
#>    73   30 0.45099809 0.06705591 0.38262948 0.13872288  0.437544556  0.12266606
#>    73   17 0.45099809 0.06705591 0.26658097 0.13426625  0.329108898  0.01922097
#>    66   65 0.80333831 0.76602569 0.70110926 0.75098393  0.743870089  0.81527949
#>    75    8 0.83667138 0.46273541 0.70047794 0.48551071  0.766949266  0.46440342
#>    64   16 0.58440755 0.09386031 0.46436137 0.17320660  0.569394134  0.20163040
#>    46   24 0.95301725 0.52534072 0.80172119 0.59931499  0.907881529  0.62473323
#>    71   96 0.72132648 0.23329331 0.79477335 0.26702483  0.768742554  0.22687695
#>    64   94 0.58440755 0.09386031 0.61786834 0.29904727  0.619124567  0.19352063
#>    84   99 0.84304143 0.46954316 0.72844298 0.52255169  0.800234097  0.52737725
#>    93   75 0.97691446 0.34501256 0.83667138 0.46273541  0.906620097  0.40366810
#>    66   33 0.80333831 0.76602569 0.71965599 0.72819923  0.750233679  0.77203030
#>    69   59 0.92640324 0.75370431 0.86546581 0.72531659  0.868688583  0.79799693
#>    96   94 0.79477335 0.26702483 0.61786834 0.29904727  0.716718624  0.34047762
#>    14   52 0.84249048 0.91510142 0.68382937 0.93845658  0.761767108  0.91731702
#>    69   66 0.92640324 0.75370431 0.80333831 0.76602569  0.868995494  0.80106234
#>    59   45 0.86546581 0.72531659 0.96681225 0.52913808  0.930367604  0.63457786
#>    93   84 0.97691446 0.34501256 0.84304143 0.46954316  0.926194716  0.42471122
#>    78   46 0.98622129 0.49765765 0.95301725 0.52534072  0.965580913  0.50665543
#>    64   81 0.58440755 0.09386031 0.78903147 0.11570666  0.687446077  0.09797806
#>    64   44 0.58440755 0.09386031 0.72156907 0.21886967  0.684957411  0.12128817
#>    22   73 0.65528720 0.00000000 0.45099809 0.06705591  0.537008051 -0.01562687
#>    22   68 0.65528720 0.00000000 0.08940749 0.14366340 -0.287804673 -2.52845908
#>    93    6 0.97691446 0.34501256 0.81485728 0.40574580  0.906380643  0.40338283
#>    80   78 1.00000000 0.77375696 0.98622129 0.49765765  1.161283643  0.62731465
#>    80   45 1.00000000 0.77375696 0.96681225 0.52913808  0.994193782  0.64998394
#>    14   47 0.84249048 0.91510142 0.99575809 0.93181886  0.910998877  0.99795497
#>    51    4 0.93378562 0.89288892 0.81287247 0.84993605  0.877413249  0.85991538
#>    87   81 0.81958807 0.10378424 0.78903147 0.11570666  0.832117406  0.18101512
#>    56   82 0.96075306 0.85830503 0.94123267 0.83160966  0.928793530  0.86119012
#>    29   83 0.82034856 0.31311546 0.90096643 0.21818239  0.857066932  0.26259979
#>    29   95 0.82034856 0.31311546 0.91863489 0.22502469  0.884803015  0.28615348
#>     6   35 0.81485728 0.40574580 0.95877349 0.31166965  0.890097214  0.36372821
#>    53   35 0.96867742 0.19097218 0.95877349 0.31166965  1.090314376  0.26170827
#>    80   50 1.00000000 0.77375696 0.96894484 0.79552953  1.019071156  0.83399291
#>    80   69 1.00000000 0.77375696 0.92640324 0.75370431  0.965051285  0.75694203
#>    83   53 0.90096643 0.21818239 0.96867742 0.19097218  0.906898148  0.13509063
#>    83   88 0.90096643 0.21818239 0.85328041 0.07133669  0.888090284  0.14119821
#>    81   88 0.78903147 0.11570666 0.85328041 0.07133669  0.763512963  0.01005304
#>    14   38 0.84249048 0.91510142 0.72624674 0.98983927  0.896873786  1.12745564
#>    47   56 0.99575809 0.93181886 0.96075306 0.85830503  0.974515636  0.89684279
#>    47   80 0.99575809 0.93181886 1.00000000 0.77375696  1.063345298  0.85454482
#>    53   93 0.96867742 0.19097218 0.97691446 0.34501256 21.497736032 -0.82954350
#>    58   28 0.23977129 0.49304972 0.23250161 0.39311987  0.248060213  0.44221737
#>     9    2 0.49390466 0.44614497 0.53215701 0.56300956  0.445599076  0.52664916
#>    55   20 0.03053718 0.47574207 0.07613862 0.44410148  0.069357571  0.48300984
#>    26   37 0.01650653 0.69806872 0.01541101 0.67686026  0.044967799  0.68596604
#>     3   90 0.26211580 0.93530732 0.10350959 0.86373467  0.188643467  0.88659991
#>    63   26 0.03866075 0.88439994 0.01650653 0.69806872 -0.093698531  0.80565442
#>    43   70 0.25156404 0.39083295 0.30623926 0.23028324  0.372970926  0.34259340
#>    36   86 0.49398500 0.86161078 0.58004779 0.84778943  0.539074003  0.86751243
#>    60   39 0.24482205 0.96788267 0.29821298 0.99007223  0.164181997  1.23724046
#>     5   63 0.08853174 0.96643892 0.03866075 0.88439994  0.006364631  0.96021017
#>    74   70 0.06634976 0.20501546 0.30623926 0.23028324  0.180621007  0.27151297
#>    13   92 0.35689790 0.16997405 0.38096979 0.21219897  0.340262711  0.20743156
#>    12   74 0.01832299 0.21990683 0.06634976 0.20501546  0.046054682  0.22445321
#>    85   10 0.50360695 0.89597181 0.55680588 0.94135654  0.485845385  0.97066314
#>    17   13 0.26658097 0.13426625 0.35689790 0.16997405  0.303768693  0.17228081
#>    33   15 0.71965599 0.72819923 0.57502159 0.70988110  0.656345118  0.64792894
#>    29    6 0.82034856 0.31311546 0.81485728 0.40574580  0.715821005  0.35339683
#>    46   59 0.95301725 0.52534072 0.86546581 0.72531659  0.907881529  0.62473323
#>    84   24 0.84304143 0.46954316 0.80172119 0.59931499  0.800234097  0.52737725
#>    38   67 0.72624674 0.98983927 0.57652391 1.00000000  0.657567890  1.08602245
#>    96   29 0.79477335 0.26702483 0.82034856 0.31311546  0.716718624  0.34047762
#>    64   22 0.58440755 0.09386031 0.65528720 0.00000000  0.687446077  0.09797806
#>    14   51 0.84249048 0.91510142 0.93378562 0.89288892  0.910998877  0.99795497
#>    87   83 0.81958807 0.10378424 0.90096643 0.21818239  0.832117406  0.18101512
#>    53   78 0.96867742 0.19097218 0.98622129 0.49765765 21.497736032 -0.82954350
#>            mx2         my2 bp1 bp2
#>    0.211831279  0.56886067   0   0
#>    0.286828092  0.44669001   0   0
#>    0.254246989  0.59624699   0   0
#>    0.335424824  0.56976513   0   0
#>    0.254246989  0.59624699   0   0
#>    0.335424824  0.56976513   0   0
#>    0.329166389  0.68206096   0   0
#>    0.328360161  0.66650121   0   0
#>    0.277506374  0.50261398   0   0
#>    0.172623531  0.50501286   0   0
#>    0.227525977  0.68493205   0   0
#>    0.172623531  0.50501286   0   0
#>    0.378017170  0.39534314   0   0
#>    0.230278572  0.29399950   0   0
#>    0.169052346  0.49206949   0   0
#>    0.476724208  0.61991216   0   0
#>    0.146480513  0.62963436   0   0
#>    0.180499110  0.57273815   0   0
#>    0.164428631  0.44830138   0   0
#>    0.081143758  0.61504838   0   0
#>    0.169052346  0.49206949   0   0
#>    0.163742335  0.44750906   0   0
#>    0.081143758  0.61504838   0   0
#>   -0.093698531  0.80565442   0   0
#>    0.139911321  0.79224938   0   0
#>    0.367139787  0.71903278   0   0
#>    0.227525977  0.68493205   0   0
#>    0.329166389  0.68206096   0   0
#>    0.490860446  0.61894988   0   0
#>    0.382268895  0.37671490   0   0
#>    0.240755365  0.81954240   0   0
#>    0.338939731  0.57148984   0   0
#>    0.512333484  0.72825028   0   0
#>    0.583533323  0.57042005   0   0
#>    0.556263948  0.46221430   0   0
#>    0.192003729  0.84552121   0   0
#>    0.475209655  0.97688281   0   0
#>    0.512333484  0.72825028   0   0
#>    0.490860446  0.61894988   0   0
#>    0.583533323  0.57042005   0   0
#>    0.577080757  0.35005557   0   0
#>    0.379042518  0.34699177   0   0
#>    0.358248304  0.80101551   0   0
#>    0.139911321  0.79224938   0   0
#>    0.177510827  0.91127003   0   0
#>    0.340262711  0.20743156   0   0
#>    0.126117877  0.33211310   0   0
#>    0.163742335  0.44750906   0   0
#>    0.277578347  0.96439431   0   0
#>    0.333725897  0.92258607   0   0
#>    0.608939995  0.52347293   0   0
#>    0.608939995  0.52347293   0   0
#>    0.608939995  0.52347293   0   0
#>    0.522179191  0.25922564   0   0
#>    0.188643467  0.88659991   0   0
#>    0.553848304  0.89095170   0   0
#>    0.060444630  0.84067665   0   0
#>   -0.366608431  5.94655871   0   1
#>    0.277578347  0.96439431   0   0
#>    0.461103691  0.27490898   0   0
#>    0.081726523  0.90746051   0   0
#>    0.006364631  0.96021017   0   0
#>    0.544240644  0.78004870   0   0
#>    0.556263948  0.46221430   0   0
#>    0.437544556  0.12266606   0   0
#>    0.230278572  0.29399950   0   0
#>   -0.149483749  0.35950848   0   0
#>    0.046054682  0.22445321   0   0
#>    0.553848304  0.89095170   0   0
#>    0.383594318  0.84861425   0   0
#>    0.682242010  0.55693191   0   0
#>    0.180834243  0.26948853   0   0
#>    0.165493215  0.20726732   0   0
#>    0.435837340  1.26827747   0   0
#>    0.719829962  0.84243203   0   0
#>    0.151036395  0.29764493   0   0
#>   -3.352680177  0.55139579   0   1
#>    0.583002496  0.96518499   0   0
#>    0.766949266  0.46440342   0   0
#>    0.577080757  0.35005557   0   0
#>    0.303342199  0.17641899   0   0
#>    0.032832167  0.18180871   0   0
#>   -0.031864025  0.26689924   0   0
#>    0.705996758  0.37153455   0   0
#>    0.325324753  0.11775835   0   0
#>    0.593816893  0.65754806   0   0
#>    0.329108898  0.01922097   0   0
#>    0.682242010  0.55693191   0   0
#>    0.705996758  0.37153455   0   0
#>    0.303768693  0.17228081   0   0
#>    0.719109033  0.85156819   0   0
#>    0.750233679  0.77203030   0   0
#>    0.297377831  4.91947453   0   1
#>    0.586491216  0.77850884   0   0
#>    0.583002496  0.96518499   0   0
#>    0.583002496  0.96518499   0   0
#>    0.698709356  0.62429272   0   0
#>    0.697515482  0.62424171   0   0
#>    0.765746939  0.45721366   0   0
#>    0.437544556  0.12266606   0   0
#>    0.175853333  0.09860044   0   0
#>    0.719829962  0.84243203   0   0
#>    0.522179191  0.25922564   0   0
#>    0.766949266  0.46440342   0   0
#>    0.793180158  0.68276103   0   0
#>    0.792107624  0.19135796   0   0
#>    0.768742554  0.22687695   0   0
#>    0.643282977  0.22476682   0   0
#>    0.698709356  0.62429272   0   0
#>    0.792433482  0.51051320   0   0
#>    0.719109033  0.85156819   0   0
#>    0.793180158  0.68276103   0   0
#>    0.604319652  0.95801740   0   0
#>    0.657567890  1.08602245   0   0
#>    0.790574148  0.68278631   0   0
#>    0.643282977  0.22476682   0   0
#>    0.715821005  0.35339683   0   0
#>    0.823747759  0.19015735   0   0
#>    0.871352326  0.55002178   0   0
#>    0.690887992  0.97582906   0   0
#>    0.868995494  0.80106234   0   0
#>    0.743870089  0.81527949   0   0
#>    0.965580913  0.50665543   0   0
#>    0.920679451  0.45279905   0   0
#>    0.994193782  0.64998394   0   0
#>    0.973692384  0.25213876   0   0
#>    0.763512963  0.01005304   0   0
#>    0.792107624  0.19135796   0   0
#>    0.511082506  0.11340840   0   0
#>    0.144823391 -0.48643640   0   0
#>    0.765746939  0.45721366   0   0
#>    0.871866576  0.80444954   0   0
#>    0.965580913  0.50665543   0   0
#>    0.832117406  0.18101512   0   0
#>    0.974515636  0.89684279   0   0
#>    0.761099349  0.91278067   0   0
#>    0.929044410  0.79356564   0   0
#>    0.871866576  0.80444954   0   0
#>    1.090314376  0.26170827   0   0
#>    0.888090284  0.14119821   0   0
#>    0.879105483  0.85515171   0   0
#>    0.823747759  0.19015735   0   0
#>    0.925437649  0.18122507   0   0
#>    0.973692384  0.25213876   0   0
#>    0.889339838  0.29121539   0   0
#>    0.974026066  0.82811483   0   0
#>    0.929044410  0.79356564   0   0
#>    5.236948729 -4.04155390   0   1
#>    2.374177042 -4.46030499   0   1
#>    1.660272695  4.67352988   0   1
#>    0.928793530  0.86119012   0   0
#>    1.019071156  0.83399291   0   0
#>    0.926194716  0.42471122   0   0
#>    0.180499110  0.57273815   0   0
#>    0.281445142  0.53965702   0   0
#>    0.211831279  0.56886067   0   0
#>    0.338939731  0.57148984   0   0
#>    0.131710271  0.75586744   0   0
#>    0.286828092  0.44669001   0   0
#>    0.416297628  0.50929678   0   0
#>    0.164705036  0.66699910   0   0
#>   -1.231290767  0.48196744   0   0
#>    0.019480517  0.57603841   0   0
#>    0.041044479  0.54443676   0   0
#>    0.037761987  0.74510634   0   0
#>    0.037761987  0.74510634   0   0
#>    0.358248304  0.80101551   0   0
#>    0.448345463  0.68549232   0   0
#>    0.448345463  0.68549232   0   0
#>    0.416297628  0.50929678   0   0
#>    0.579592912  0.56952308   0   0
#>    0.192003729  0.84552121   0   0
#>    0.579592912  0.56952308   0   0
#>    0.384399556  0.85938935   0   0
#>    0.593816893  0.65754806   0   0
#>    0.697515482  0.62424171   0   0
#>    0.491652447  0.40162008   0   0
#>    0.491652447  0.40162008   0   0
#>    0.060444630  0.84067665   0   0
#>    0.166724556  0.96200169   0   0
#>    0.379042518  0.34699177   0   0
#>   -0.014549717  0.36207996   0   0
#>    0.383594318  0.84861425   0   0
#>    0.596722385  0.44719551   0   0
#>    0.425512132  0.89931262   0   0
#>    0.169839424  0.92167891   0   0
#>    0.166724556  0.96200169   0   0
#>    0.333725897  0.92258607   0   0
#>    0.627974169  0.76138862   0   0
#>    0.637863179  0.39872043   0   0
#>    0.461103691  0.27490898   0   0
#>    0.151036395  0.29764493   0   0
#>   -0.149483749  0.35950848   0   0
#>    0.620594941  0.90556025   0   0
#>    0.544240644  0.78004870   0   0
#>    0.608822571  0.48579795   0   0
#>    0.303342199  0.17641899   0   0
#>   -0.149483749  0.35950848   0   0
#>    0.475209655  0.97688281   0   0
#>    0.682477048  0.53666964   0   0
#>    0.648890014  0.39683476   0   0
#>    0.032477218  0.15727638   0   0
#>   -5.214194178 -7.37083032   0   1
#>    0.685823263  0.52565160   0   0
#>    0.461103691  0.27490898   0   0
#>    0.414953766  0.17620981   0   0
#>    0.659185558  0.64398660   0   0
#>    0.165493215  0.20726732   0   0
#>    0.690887992  0.97582906   0   0
#>    0.627974169  0.76138862   0   0
#>    0.332200797  0.92839104   0   0
#>    0.621968605  0.90528139   0   0
#>    0.425386344  1.33093726   0   0
#>    0.719829962  0.84243203   0   0
#>    0.618306894  0.90696311   0   0
#>    0.329108898  0.01922097   0   0
#>    0.144823391 -0.48643640   0   0
#>    0.750233679  0.77203030   0   0
#>    0.765746939  0.45721366   0   0
#>    0.511082506  0.11340840   0   0
#>    0.871352326  0.55002178   0   0
#>    0.716676562  0.34024525   0   0
#>    0.569394134  0.20163040   0   0
#>    0.792433482  0.51051320   0   0
#>    0.906380643  0.40338283   0   0
#>    0.790574148  0.68278631   0   0
#>    0.942014467  0.64059467   0   0
#>    0.716676562  0.34024525   0   0
#>    0.761099349  0.91278067   0   0
#>    0.868688583  0.79799693   0   0
#>    0.942014467  0.64059467   0   0
#>    0.906620097  0.40366810   0   0
#>    0.920679451  0.45279905   0   0
#>    0.684957411  0.12128817   0   0
#>    0.619124567  0.19352063   0   0
#>   -0.287804673 -2.52845908   0   0
#>   -0.709279244 -4.18861675   0   1
#>    0.893430520  0.36882745   0   0
#>    4.774170928  0.44701390   0   1
#>    1.161283643  0.62731465   0   0
#>    0.896873786  1.12745564   0   0
#>    0.879105483  0.85515171   0   0
#>    0.767879075  0.01637529   0   0
#>    0.974026066  0.82811483   0   0
#>    0.884803015  0.28615348   0   0
#>    0.889339838  0.29121539   0   0
#>    0.893430520  0.36882745   0   0
#>    0.973692384  0.25213876   0   0
#>    0.965051285  0.75694203   0   0
#>    0.994193782  0.64998394   0   0
#>    0.925437649  0.18122507   0   0
#>    0.906898148  0.13509063   0   0
#>    0.767879075  0.01637529   0   0
#>    0.761767108  0.91731702   0   0
#>    1.063345298  0.85454482   0   0
#>    7.385427325  1.02421053   0   1
#>    1.090314376  0.26170827   0   0
#>    0.164428631  0.44830138   0   0
#>    0.522117759  0.50160292   0   0
#>   -0.014549717  0.36207996   0   0
#>  -46.980548845  3.11505057   0   1
#>    0.177510827  0.91127003   0   0
#>   -5.385675812  1.43485470   0   1
#>    0.230278572  0.29399950   0   0
#>    0.528884673  0.80406551   0   0
#>    0.277578347  0.96439431   0   0
#>   -8.893976075  6.37067222   0   1
#>    0.180834243  0.26948853   0   0
#>    0.395788679  0.17577691   0   0
#>    0.032832167  0.18180871   0   0
#>    0.553848304  0.89095170   0   0
#>    0.325324753  0.11775835   0   0
#>    0.650871200  0.69114934   0   0
#>    0.890097214  0.36372821   0   0
#>    0.930367604  0.63457786   0   0
#>    0.871352326  0.55002178   0   0
#>    1.108751718  7.73441178   0   1
#>    0.857066932  0.26259979   0   0
#>    0.537008051 -0.01562687   0   0
#>    0.877413249  0.85991538   0   0
#>    0.888090284  0.14119821   0   0
#>   24.747770316 -1.01546098   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.27313740 0.59377765
#>   [2,] 0.53215701 0.56300956
#>   [3,] 0.26211580 0.93530732
#>   [4,] 0.81287247 0.84993605
#>   [5,] 0.08853174 0.96643892
#>   [6,] 0.81485728 0.40574580
#>   [7,] 0.62676203 0.83525018
#>   [8,] 0.70047794 0.48551071
#>   [9,] 0.49390466 0.44614497
#>  [10,] 0.55680588 0.94135654
#>  [11,] 0.46772692 0.77597324
#>  [12,] 0.01832299 0.21990683
#>  [13,] 0.35689790 0.16997405
#>  [14,] 0.84249048 0.91510142
#>  [15,] 0.57502159 0.70988110
#>  [16,] 0.46436137 0.17320660
#>  [17,] 0.26658097 0.13426625
#>  [18,] 0.31630352 0.50580654
#>  [19,] 0.53935681 0.66877789
#>  [20,] 0.07613862 0.44410148
#>  [21,] 0.65369049 0.96789982
#>  [22,] 0.65528720 0.00000000
#>  [23,] 0.51295881 0.36245916
#>  [24,] 0.80172119 0.59931499
#>  [25,] 0.08865039 0.95581866
#>  [26,] 0.01650653 0.69806872
#>  [27,] 0.27095963 0.58710189
#>  [28,] 0.23250161 0.39311987
#>  [29,] 0.82034856 0.31311546
#>  [30,] 0.38262948 0.13872288
#>  [31,] 0.10009925 0.50780485
#>  [32,] 0.39822756 0.60770651
#>  [33,] 0.71965599 0.72819923
#>  [34,] 0.53623014 0.40101340
#>  [35,] 0.95877349 0.31166965
#>  [36,] 0.49398500 0.86161078
#>  [37,] 0.01541101 0.67686026
#>  [38,] 0.72624674 0.98983927
#>  [39,] 0.29821298 0.99007223
#>  [40,] 0.25222679 0.15918632
#>  [41,] 0.58101720 0.52166327
#>  [42,] 0.62572718 0.54585925
#>  [43,] 0.25156404 0.39083295
#>  [44,] 0.72156907 0.21886967
#>  [45,] 0.96681225 0.52913808
#>  [46,] 0.95301725 0.52534072
#>  [47,] 0.99575809 0.93181886
#>  [48,] 0.38983595 0.96886240
#>  [49,] 0.55965052 0.99180720
#>  [50,] 0.96894484 0.79552953
#>  [51,] 0.93378562 0.89288892
#>  [52,] 0.68382937 0.93845658
#>  [53,] 0.96867742 0.19097218
#>  [54,] 0.24024803 0.51389348
#>  [55,] 0.03053718 0.47574207
#>  [56,] 0.96075306 0.85830503
#>  [57,] 0.63685096 0.52148925
#>  [58,] 0.23977129 0.49304972
#>  [59,] 0.86546581 0.72531659
#>  [60,] 0.24482205 0.96788267
#>  [61,] 0.27636151 0.87787407
#>  [62,] 0.25668405 0.75308446
#>  [63,] 0.03866075 0.88439994
#>  [64,] 0.58440755 0.09386031
#>  [65,] 0.70110926 0.75098393
#>  [66,] 0.80333831 0.76602569
#>  [67,] 0.57652391 1.00000000
#>  [68,] 0.08940749 0.14366340
#>  [69,] 0.92640324 0.75370431
#>  [70,] 0.30623926 0.23028324
#>  [71,] 0.72132648 0.23329331
#>  [72,] 0.06969028 0.70454877
#>  [73,] 0.45099809 0.06705591
#>  [74,] 0.06634976 0.20501546
#>  [75,] 0.83667138 0.46273541
#>  [76,] 0.03483506 0.25021822
#>  [77,] 0.10884141 0.52917438
#>  [78,] 0.98622129 0.49765765
#>  [79,] 0.05969249 0.79183304
#>  [80,] 1.00000000 0.77375696
#>  [81,] 0.78903147 0.11570666
#>  [82,] 0.94123267 0.83160966
#>  [83,] 0.90096643 0.21818239
#>  [84,] 0.84304143 0.46954316
#>  [85,] 0.50360695 0.89597181
#>  [86,] 0.58004779 0.84778943
#>  [87,] 0.81958807 0.10378424
#>  [88,] 0.85328041 0.07133669
#>  [89,] 0.65212562 0.49937387
#>  [90,] 0.10350959 0.86373467
#>  [91,] 0.56016141 0.61328388
#>  [92,] 0.38096979 0.21219897
#>  [93,] 0.97691446 0.34501256
#>  [94,] 0.61786834 0.29904727
#>  [95,] 0.91863489 0.22502469
#>  [96,] 0.79477335 0.26702483
#>  [97,] 0.00000000 0.20597545
#>  [98,] 0.21219355 0.75745695
#>  [99,] 0.72844298 0.52255169
#> [100,] 0.25279034 0.61524234
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,100,27), [2,4,6]
#> 2: (54,27,100), [1,11,5]
#> 3: (54,58,18), [14,5,10]
#> 4: (18,1,27), [1,5,7]
#> 5: (54,18,27), [4,2,3]
#> 6: (32,100,1), [1,7,9]
#> 7: (32,1,18), [4,33,6]
#> 8: (98,100,62), [9,28,12]
#> 9: (62,100,32), [6,29,8]
#> 10: (77,58,54), [3,11,13]
#> 11: (100,77,54), [10,2,19]
#> 12: (72,100,98), [8,26,19]
#> 13: (31,58,77), [10,22,16]
#> 14: (58,43,18), [31,3,15]
#> 15: (58,28,43), [53,14,16]
#> 16: (28,58,31), [13,20,15]
#> 17: (9,2,32), [30,33,38]
#> 18: (37,77,72), [19,24,21]
#> 19: (72,77,100), [11,12,18]
#> 20: (20,28,31), [16,23,55]
#> 21: (55,77,37), [18,88,22]
#> 22: (55,31,77), [13,21,23]
#> 23: (55,20,31), [20,22,76]
#> 24: (26,37,72), [18,25,0]
#> 25: (72,79,26), [49,24,26]
#> 26: (72,98,79), [37,25,12]
#> 27: (61,62,11), [29,48,28]
#> 28: (61,98,62), [8,27,32]
#> 29: (11,62,32), [9,34,27]
#> 30: (32,2,19), [35,34,17]
#> 31: (18,43,9), [47,33,14]
#> 32: (90,98,61), [28,39,37]
#> 33: (9,32,18), [7,31,17]
#> 34: (32,19,11), [42,29,30]
#> 35: (91,19,2), [30,36,43]
#> 36: (2,41,91), [58,35,38]
#> 37: (90,79,98), [26,32,50]
#> 38: (2,9,41), [45,36,17]
#> 39: (3,90,61), [32,63,56]
#> 40: (36,85,48), [85,57,64]
#> 41: (86,11,15), [42,72,79]
#> 42: (15,11,19), [34,43,41]
#> 43: (15,19,91), [35,102,42]
#> 44: (33,91,42), [58,104,102]
#> 45: (34,41,9), [38,46,73]
#> 46: (9,23,34), [62,45,47]
#> 47: (9,43,23), [69,46,31]
#> 48: (36,61,11), [27,79,57]
#> 49: (63,26,79), [25,50,0]
#> 50: (63,79,90), [37,65,49]
#> 51: (25,90,60), [56,71,65]
#> 52: (43,70,92), [84,69,53]
#> 53: (70,43,28), [15,75,52]
#> 54: (74,28,76), [55,87,75]
#> 55: (76,28,20), [20,76,54]
#> 56: (90,3,60), [67,51,39]
#> 57: (36,48,61), [63,48,40]
#> 58: (91,41,42), [59,44,36]
#> 59: (57,42,41), [58,60,61]
#> 60: (89,57,41), [59,73,80]
#> 61: (99,42,57), [59,80,104]
#> 62: (34,23,94), [99,92,46]
#> 63: (48,3,61), [39,57,68]
#> 64: (36,86,85), [90,40,79]
#> 65: (25,63,90), [50,51,70]
#> 66: (60,39,5), [0,71,67]
#> 67: (39,60,3), [56,68,66]
#> 68: (48,39,3), [67,63,109]
#> 69: (43,92,23), [83,47,52]
#> 70: (5,63,25), [65,71,0]
#> 71: (60,5,25), [70,51,66]
#> 72: (7,86,15), [41,86,110]
#> 73: (89,41,34), [45,92,60]
#> 74: (92,30,16), [119,83,100]
#> 75: (74,70,28), [53,54,82]
#> 76: (20,55,76), [97,55,23]
#> 77: (97,12,76), [87,97,96]
#> 78: (21,10,86), [90,110,113]
#> 79: (86,36,11), [48,41,64]
#> 80: (89,99,57), [61,60,91]
#> 81: (13,70,40), [82,93,84]
#> 82: (70,74,40), [94,81,75]
#> 83: (16,23,92), [69,74,99]
#> 84: (13,92,70), [52,81,100]
#> 85: (85,49,48), [111,40,89]
#> 86: (15,65,7), [117,72,108]
#> 87: (12,74,76), [54,77,96]
#> 88: (55,37,97), [0,97,21]
#> 89: (85,10,49), [101,85,90]
#> 90: (10,85,86), [64,78,89]
#> 91: (89,8,99), [123,80,98]
#> 92: (89,34,94), [62,98,73]
#> 93: (17,13,40), [81,106,103]
#> 94: (68,40,74), [82,95,106]
#> 95: (68,74,97), [96,0,94]
#> 96: (74,12,97), [77,95,87]
#> 97: (97,76,55), [76,88,77]
#> 98: (89,94,8), [116,91,92]
#> 99: (16,94,23), [62,83,124]
#> 100: (92,13,30), [103,74,84]
#> 101: (67,49,10), [89,113,111]
#> 102: (33,15,91), [43,44,108]
#> 103: (13,17,30), [120,100,93]
#> 104: (33,42,99), [61,115,44]
#> 105: (29,6,94), [116,140,172]
#> 106: (68,17,40), [93,94,121]
#> 107: (21,7,52), [112,137,110]
#> 108: (15,33,65), [134,86,102]
#> 109: (48,67,39), [0,68,111]
#> 110: (21,86,7), [72,107,78]
#> 111: (48,49,67), [101,109,85]
#> 112: (4,52,7), [107,117,133]
#> 113: (21,67,10), [101,78,136]
#> 114: (66,33,24), [115,138,134]
#> 115: (24,33,99), [104,131,114]
#> 116: (94,6,8), [126,98,105]
#> 117: (4,7,65), [86,122,112]
#> 118: (64,16,73), [119,154,124]
#> 119: (73,16,30), [74,120,118]
#> 120: (73,30,17), [103,121,119]
#> 121: (73,17,68), [106,155,120]
#> 122: (66,4,65), [117,134,145]
#> 123: (75,99,8), [91,126,125]
#> 124: (64,94,16), [99,118,130]
#> 125: (84,99,75), [123,132,131]
#> 126: (75,8,6), [116,156,123]
#> 127: (46,59,24), [138,147,146]
#> 128: (71,44,96), [141,139,129]
#> 129: (94,44,71), [128,139,130]
#> 130: (64,44,94), [129,124,153]
#> 131: (84,24,99), [115,125,147]
#> 132: (93,84,75), [125,156,148]
#> 133: (14,52,4), [112,162,143]
#> 134: (66,65,33), [108,114,122]
#> 135: (69,66,59), [138,150,145]
#> 136: (38,67,21), [113,137,0]
#> 137: (52,38,21), [136,107,143]
#> 138: (59,66,24), [114,127,135]
#> 139: (96,94,71), [129,128,140]
#> 140: (96,29,94), [105,139,169]
#> 141: (44,81,96), [160,128,153]
#> 142: (78,46,84), [147,148,149]
#> 143: (14,38,52), [137,133,179]
#> 144: (82,4,69), [145,164,157]
#> 145: (69,4,66), [122,135,144]
#> 146: (59,46,45), [149,150,127]
#> 147: (24,84,46), [142,127,131]
#> 148: (93,78,84), [142,132,182]
#> 149: (78,45,46), [146,142,158]
#> 150: (59,45,69), [159,135,146]
#> 151: (29,95,35), [183,172,170]
#> 152: (64,22,81), [178,153,154]
#> 153: (64,81,44), [141,130,152]
#> 154: (22,64,73), [118,155,152]
#> 155: (22,73,68), [121,0,154]
#> 156: (93,75,6), [126,165,132]
#> 157: (51,4,82), [144,168,162]
#> 158: (80,45,78), [149,0,159]
#> 159: (80,69,45), [150,158,175]
#> 160: (96,81,83), [166,169,141]
#> 161: (14,51,47), [180,179,162]
#> 162: (51,14,4), [133,157,161]
#> 163: (56,82,50), [164,174,168]
#> 164: (50,82,69), [144,175,163]
#> 165: (6,35,93), [173,156,172]
#> 166: (87,83,81), [160,167,177]
#> 167: (81,88,87), [177,166,178]
#> 168: (56,51,82), [157,163,180]
#> 169: (29,96,83), [160,170,140]
#> 170: (29,83,95), [171,151,169]
#> 171: (83,53,95), [183,170,176]
#> 172: (6,29,35), [151,165,105]
#> 173: (53,93,35), [165,183,182]
#> 174: (80,56,50), [163,175,181]
#> 175: (80,50,69), [164,159,174]
#> 176: (83,88,53), [0,171,177]
#> 177: (83,87,88), [167,176,166]
#> 178: (81,22,88), [0,167,152]
#> 179: (14,47,38), [0,143,161]
#> 180: (47,51,56), [168,181,161]
#> 181: (47,56,80), [174,0,180]
#> 182: (53,78,93), [148,173,0]
#> 183: (53,35,95), [151,171,173]
#> boundary nodes:  26 37 97 68 22 88 53 78 80 47 38 67 39 5 63 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.033496390 0.077588134 0.007021999 0.093085995 0.029575933 0.097990972 
#>                                                                         
#> 0.137897102 0.145632322 0.020849212 0.132292129 0.147895549 0.135822041 
#>                                                                         
#> 0.131947441 0.019199113 0.140449254 0.141191066 0.179691365 0.167716990 
#>                                                                         
#> 0.175165721 0.174757993 0.023088577 0.068060490 0.060933511 0.103231661 
#>                                                                         
#> 0.156327506 0.212280458 0.044704840 0.202902227 0.106013099 0.248572723 
#>                                                                         
#> 0.136447094 0.130748427 0.128925106 0.093964358 0.115289233 0.173429265 
#>                                                                         
#> 0.135118024 0.126017126 0.059265644 0.094047600 0.045033145 0.262930217 
#>                                                                         
#> 0.216805211 0.084200846 0.175544919 0.076887536 0.243911763 0.164464319 
#>                                                                         
#> 0.036881256 0.145448700 0.050837271 0.055834034 0.026788724 0.122584985 
#>                                                                         
#> 0.059173637 0.090358970 0.068061921 0.211008895 0.065591168 0.199998060 
#>                                                                         
#> 0.087175677 0.010620917 0.137999892 0.128694492 0.088708672 0.178754043 
#>                                                                         
#> 0.225564800 0.034517082 0.096410532 0.089572737 0.091598182 0.089286719 
#>                                                                         
#> 0.191443408 0.171357676 0.112375761 0.055104139 0.471136922 0.050530796 
#>                                                                         
#> 0.046412050 0.130621161 0.105225550 0.066356707 0.023017716 0.203943524 
#>                                                                         
#> 0.040481480 0.097733559 0.116134052 0.105327006 0.224029506 0.028758522 
#>                                                                         
#> 0.117933207 0.029378970 0.278487947 0.048367889 0.018757217 0.061869668 
#>                                                                         
#> 0.152793471 0.205835176 0.139445600 0.106988530 0.177422508 0.149273359 
#>                                                                         
#> 0.198495002 0.123658294 0.141208292 0.087622977 0.014425687 0.131081229 
#>                                                                         
#> 0.106123974 0.009323254 0.156486446 0.074276893 0.083576929 0.075800707 
#>                                                                         
#> 0.166718547 0.122585349 0.202968358 0.151427074 0.123320915 0.066628927 
#>                                                                         
#> 0.148828031 0.084450273 0.014308101 0.145914005 0.228172916 0.095490613 
#>                                                                         
#> 0.176848975 0.123263058 0.136075562 0.369616652 0.061021883 0.129661850 
#>                                                                         
#> 0.036982815 0.151758731 0.073185566 0.071580392 0.045494406 0.079304201 
#>                                                                         
#> 0.037958453 0.046776244 0.061730107 0.116886930 0.018947071 0.060529565 
#>                                                                         
#> 0.138432479 0.063307738 0.059658504 0.166220094 0.210452456 0.275685944 
#>                                                                         
#> 0.043855308 0.093213194 0.152928549 0.102121988 0.076484223 0.079389375 
#>                                                                         
#> 0.125863263 0.152008085 0.102894785 0.187766449 0.203718604 0.201686206 
#>                                                                         
#> 0.094797485 0.076595710 0.053577073 0.087854993 0.126331501 0.182054533 
#>                                                                         
#> 0.153776389 0.187354389 0.057547837 0.152010388 0.064006513 0.149498944 
#>                                                                         
#> 0.054419215 0.205110916 0.061873300 0.085827595 0.094926096 0.156636936 
#>                                                                         
#> 0.220580934 0.198233962 0.218230324 0.074519959 0.035682796 0.093275170 
#>                                                                         
#> 0.156296982 0.132054455 0.135626247 0.152008371 0.092057374 0.250977496 
#>                                                                         
#> 0.056310778 0.140889156 0.133317451 0.026877593 0.078762274 0.271489500 
#>                                                                         
#> 0.111019391 0.079759337 0.203234615 0.065541850 0.108979317 0.050300434 
#>                                                                         
#> 0.195392536 0.073494839 0.196580926 0.163557583 0.042133792 0.132618039 
#>                                                                         
#> 0.094045870 0.135355349 0.189266863 0.186688972 0.100454835 0.099047572 
#>                                                                         
#> 0.196282719 0.103329727 0.138084643 0.143898992 0.168412267 0.080822388 
#>                                                                         
#> 0.207897360 0.126264442 0.183103230 0.091834480 0.067225237 0.179779916 
#>                                                                         
#> 0.160370862 0.123680204 0.220810120 0.182838343 0.043230322 0.205786807 
#>                                                                         
#> 0.185581849 0.215012865 0.583831320 0.173063740 0.276442909 0.246859929 
#>                                                                         
#> 0.154176629 0.128315778 0.032800153 0.033070966 0.124545286 0.131985553 
#>                                                                         
#> 0.171936610 0.121103129 0.037927135 0.076279696 0.072973788 0.154394351 
#>                                                                         
#> 0.078080859 0.138196795 0.081422567 0.158118809 0.154260451 0.100193936 
#>                                                                         
#> 0.122965739 0.055503310 0.021236738 0.174007396 0.187643628 0.169604220 
#>                                                                         
#> 0.087165560 0.057818396 0.096007866 0.241216562 0.048604527 0.050282435 
#>                                                                         
#> 0.069927824 0.097119487 0.145789792 0.092792966 0.218301637 0.136191371 
#>                                                                         
#> 0.150067202 0.052710889 0.117616673 0.093958489 0.140390083 0.307186855 
#> 
#> $alpha
#> [1] 0.301054
#> 
#> attr(,"class")
#> [1] "scree"