Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2           x1          y1           x2          y2         mx1
#>    73   18 0.5150122847 0.388362764 0.4100521792 0.372344619  0.46371971
#>     2   22 0.5278121866 0.422022382 0.3939444383 0.415222653  0.46038162
#>    22   18 0.3939444383 0.415222653 0.4100521792 0.372344619  0.45731152
#>     1   18 0.4466615437 0.321687757 0.4100521792 0.372344619  0.38150523
#>    38   62 0.2510969860 0.384233831 0.3446781846 0.258733578  0.32800552
#>    22   30 0.3939444383 0.415222653 0.2763887113 0.422632833  0.33192219
#>    77   30 0.2450172421 0.448829346 0.2763887113 0.422632833  0.33790894
#>    36   62 0.3160308839 0.244458235 0.3446781846 0.258733578  0.29616759
#>    57    2 0.6848346683 0.404749044 0.5278121866 0.422022382  0.60204892
#>    73   25 0.5150122847 0.388362764 0.5320513438 0.179344111  0.54869008
#>    73    1 0.5150122847 0.388362764 0.4466615437 0.321687757  0.54829468
#>     1   62 0.4466615437 0.321687757 0.3446781846 0.258733578  0.43886920
#>    25   62 0.5320513438 0.179344111 0.3446781846 0.258733578  0.40601691
#>    87   50 0.6521022497 0.478622504 0.5369571944 0.465667446  0.59929107
#>    41   38 0.2587567555 0.429048097 0.2510969860 0.384233831  0.26025573
#>    50   15 0.5369571944 0.465667446 0.4678868140 0.592247867  0.44523643
#>    30   38 0.2763887113 0.422632833 0.2510969860 0.384233831  0.32922909
#>    91   57 0.6620263304 0.308433610 0.6848346683 0.404749044  0.60215096
#>    77   41 0.2450172421 0.448829346 0.2587567555 0.429048097  0.27925468
#>    91   73 0.6620263304 0.308433610 0.5150122847 0.388362764  0.56125767
#>    57   87 0.6848346683 0.404749044 0.6521022497 0.478622504  0.60641182
#>    51   26 0.8159650173 0.295102158 0.8111563873 0.452032598  0.76902730
#>    15   89 0.4678868140 0.592247867 0.2705497243 0.630117475  0.35470738
#>     6   54 0.2344680082 0.461534878 0.1929083827 0.401929694  0.21249858
#>    71   97 0.6505488198 0.250454038 0.6395240227 0.188045164  0.66779256
#>    71   25 0.6505488198 0.250454038 0.5320513438 0.179344111  0.58200769
#>    97   25 0.6395240227 0.188045164 0.5320513438 0.179344111  0.58666378
#>     7   62 0.5188617954 0.075654331 0.3446781846 0.258733578  0.40252441
#>    26   64 0.8111563873 0.452032598 0.7403136379 0.608850651  0.75890986
#>    54   48 0.1929083827 0.401929694 0.1600454796 0.337072368  0.20971047
#>    97    9 0.6395240227 0.188045164 0.5503721815 0.131553705  0.58866220
#>     7    5 0.5188617954 0.075654331 0.5815826525 0.116122553  0.55131844
#>    51   91 0.8159650173 0.295102158 0.6620263304 0.308433610  0.73773322
#>   100   67 0.6301912396 0.105026970 0.7070090079 0.214810537  0.69784569
#>     6   69 0.2344680082 0.461534878 0.2015611267 0.500415584  0.20511918
#>    89   69 0.2705497243 0.630117475 0.2015611267 0.500415584  0.28613488
#>     9   25 0.5503721815 0.131553705 0.5320513438 0.179344111  0.48254329
#>    14   58 0.1847591012 0.071820591 0.2711813997 0.098344998  0.19576289
#>    48   36 0.1600454796 0.337072368 0.3160308839 0.244458235  0.24003161
#>     5  100 0.5815826525 0.116122553 0.6301912396 0.105026970  0.61461506
#>    89   20 0.2705497243 0.630117475 0.2171869312 0.619036821  0.25868323
#>     6   77 0.2344680082 0.461534878 0.2450172421 0.448829346  0.32915653
#>    15   99 0.4678868140 0.592247867 0.5702165733 0.681418154  0.47157659
#>    11   89 0.3199489062 0.674808412 0.2705497243 0.630117475  0.36280925
#>     4   16 0.3541497732 0.014432560 0.4069079741 0.006799476  0.39903218
#>    53   23 0.1764550764 0.576789613 0.1781907247 0.484514858  0.16406397
#>    53   49 0.1764550764 0.576789613 0.0496714908 0.423565645  0.07848020
#>    20   53 0.2171869312 0.619036821 0.1764550764 0.576789613  0.24071012
#>    64   87 0.7403136379 0.608850651 0.6521022497 0.478622504  0.71427960
#>    26   80 0.8111563873 0.452032598 0.7095124588 0.450016374  0.76149899
#>    91   71 0.6620263304 0.308433610 0.6505488198 0.250454038  0.70254974
#>    29   51 0.8312221603 0.249472198 0.8159650173 0.295102158  0.76351823
#>    99   87 0.5702165733 0.681418154 0.6521022497 0.478622504  0.63088365
#>    16   96 0.4069079741 0.006799476 0.6682576706 0.021180060  0.54327644
#>     4   62 0.3541497732 0.014432560 0.3446781846 0.258733578  0.39558271
#>    47   67 0.8480127753 0.101363319 0.7070090079 0.214810537  0.73922489
#>    11   15 0.3199489062 0.674808412 0.4678868140 0.592247867  0.43426573
#>    99   17 0.5702165733 0.681418154 0.6839011949 0.726409285  0.62068958
#>    26   92 0.8111563873 0.452032598 0.9532668571 0.428360448  0.88812130
#>    82   26 0.8654626457 0.380283500 0.8111563873 0.452032598  0.78070902
#>    51   75 0.8159650173 0.295102158 0.8639968879 0.304961370  0.83331615
#>    34   29 0.8635004214 0.166238918 0.8312221603 0.249472198  0.78287143
#>    96  100 0.6682576706 0.021180060 0.6301912396 0.105026970  0.58893426
#>    14   44 0.1847591012 0.071820591 0.2590521812 0.014244331  0.23549459
#>    58   36 0.2711813997 0.098344998 0.3160308839 0.244458235  0.19853040
#>    82   51 0.8654626457 0.380283500 0.8159650173 0.295102158  0.82744947
#>    44   95 0.2590521812 0.014244331 0.2770127215 0.000000000  0.30272859
#>    95    4 0.2770127215 0.000000000 0.3541497732 0.014432560  0.30736207
#>    95   16 0.2770127215 0.000000000 0.4069079741 0.006799476  0.35192982
#>    64   17 0.7403136379 0.608850651 0.6839011949 0.726409285  0.65270027
#>    64   52 0.7403136379 0.608850651 0.8080388551 0.623736844  0.76053395
#>    48   49 0.1600454796 0.337072368 0.0496714908 0.423565645  0.07656883
#>    54   23 0.1929083827 0.401929694 0.1781907247 0.484514858  0.12424282
#>    81   99 0.5357936883 0.766838802 0.5702165733 0.681418154  0.57545581
#>    33   84 0.0850819863 0.147099916 0.1093853138 0.069653514  0.14555406
#>    74   99 0.5788572291 0.785086372 0.5702165733 0.681418154  0.61000122
#>    78   19 0.7851591563 0.746433158 0.6573278026 0.771914724  0.72565469
#>    48   24 0.1600454796 0.337072368 0.0169148210 0.403047325  0.07001749
#>    49   40 0.0496714908 0.423565645 0.1040421188 0.741456118 -0.03381158
#>    29    3 0.8312221603 0.249472198 0.8881957112 0.167826981  0.87377085
#>    34    3 0.8635004214 0.166238918 0.8881957112 0.167826981  0.87259366
#>    96   47 0.6682576706 0.021180060 0.8480127753 0.101363319  0.73911409
#>    81   27 0.5357936883 0.766838802 0.5333868682 0.821445378  0.47912959
#>    86   27 0.5412423298 0.817867353 0.5333868682 0.821445378  0.52551102
#>    81   74 0.5357936883 0.766838802 0.5788572291 0.785086372  0.55094236
#>    74   19 0.5788572291 0.785086372 0.6573278026 0.771914724  0.61896290
#>    11   37 0.3199489062 0.674808412 0.4485215172 0.845474257  0.36370771
#>    86   94 0.5412423298 0.817867353 0.6035318922 0.820730584  0.56950931
#>    86   74 0.5412423298 0.817867353 0.5788572291 0.785086372  0.57254716
#>    47   34 0.8480127753 0.101363319 0.8635004214 0.166238918  0.77372699
#>    68   46 0.9383035630 0.287289448 0.8729073784 0.320999731  0.90087481
#>    11   81 0.3199489062 0.674808412 0.5357936883 0.766838802  0.42382351
#>    32   37 0.5022711789 0.862439104 0.4485215172 0.845474257  0.48665650
#>    33   48 0.0850819863 0.147099916 0.1600454796 0.337072368  0.10239487
#>    28   68 0.9206979757 0.242333838 0.9383035630 0.287289448  0.89675308
#>    75   29 0.8639968879 0.304961370 0.8312221603 0.249472198  0.87707004
#>    82   10 0.8654626457 0.380283500 0.9555826481 0.382968668  0.90987413
#>    40   53 0.1040421188 0.741456118 0.1764550764 0.576789613  0.13550005
#>    12   40 0.2622078278 0.817709859 0.1040421188 0.741456118  0.20931572
#>    37   63 0.4485215172 0.845474257 0.4262983784 0.967513254  0.38552366
#>    82   46 0.8654626457 0.380283500 0.8729073784 0.320999731  0.91128831
#>    82   92 0.8654626457 0.380283500 0.9532668571 0.428360448  0.88396343
#>    94   19 0.6035318922 0.820730584 0.6573278026 0.771914724  0.68840056
#>    78   93 0.7851591563 0.746433158 0.7936747403 0.828611652  0.72815971
#>    83   78 0.8157843325 0.726831319 0.7851591563 0.746433158  0.76353192
#>    52   26 0.8080388551 0.623736844 0.8111563873 0.452032598  0.86132505
#>    59   85 0.9445794478 0.205076860 1.0000000000 0.367663354  1.06296948
#>    28    3 0.9206979757 0.242333838 0.8881957112 0.167826981  0.90393243
#>    79   59 0.9465931931 0.102012070 0.9445794478 0.205076860  0.93822292
#>    88   83 0.8972254874 0.656380294 0.8157843325 0.726831319  0.84052604
#>    85   61 1.0000000000 0.367663354 0.9586015217 0.514597579  1.06341701
#>    61   88 0.9586015217 0.514597579 0.8972254874 0.656380294  0.88016996
#>    45   43 0.3053016205 0.883162585 0.2345564400 0.869332644  0.27367791
#>    45   12 0.3053016205 0.883162585 0.2622078278 0.817709859  0.36004129
#>    10   85 0.9555826481 0.382968668 1.0000000000 0.367663354  0.98885407
#>    43   40 0.2345564400 0.869332644 0.1040421188 0.741456118  0.17201944
#>    47   79 0.8480127753 0.101363319 0.9465931931 0.102012070  0.89720220
#>    13   79 0.9054318963 0.017935214 0.9465931931 0.102012070  0.89748559
#>    39   32 0.5568987646 0.952751682 0.5022711789 0.862439104  0.57001736
#>    96   13 0.6682576706 0.021180060 0.9054318963 0.017935214  0.78654416
#>    63   98 0.4262983784 0.967513254 0.2986739179 0.918012248  0.37356467
#>    32   70 0.5022711789 0.862439104 0.5111976552 0.949869835  0.45884605
#>    39   94 0.5568987646 0.952751682 0.6035318922 0.820730584  0.63156418
#>    39   70 0.5568987646 0.952751682 0.5111976552 0.949869835  0.53709114
#>    35   31 0.8462844438 0.827614859 0.7969770907 0.850551389  0.82012615
#>    31   42 0.7969770907 0.850551389 0.7067057662 0.952000767  0.70161527
#>    35   88 0.8462844438 0.827614859 0.8972254874 0.656380294  0.95245492
#>    32   63 0.5022711789 0.862439104 0.4262983784 0.967513254  0.45765087
#>    21   42 0.8015995365 0.879146532 0.7067057662 0.952000767  0.72449180
#>    65   21 0.8340500675 0.863276476 0.8015995365 0.879146532  0.81358309
#>    35   65 0.8462844438 0.827614859 0.8340500675 0.863276476  0.82163257
#>    76   65 0.9050062503 0.895773977 0.8340500675 0.863276476  0.87903013
#>     8   24 0.0004357169 0.272443903 0.0169148210 0.403047325 -1.53436718
#>    24   49 0.0169148210 0.403047325 0.0496714908 0.423565645 -0.08062880
#>    72   40 0.0000000000 0.794998361 0.1040421188 0.741456118  0.09867455
#>    42   56 0.7067057662 0.952000767 0.7612457790 0.991453475  0.63260173
#>    93   31 0.7936747403 0.828611652 0.7969770907 0.850551389  0.70339780
#>    21   76 0.8015995365 0.879146532 0.9050062503 0.895773977  0.84434769
#>    98   66 0.2986739179 0.918012248 0.1575476249 0.960627633  0.22920564
#>    76   35 0.9050062503 0.895773977 0.8462844438 0.827614859  0.90973000
#>    61   60 0.9586015217 0.514597579 0.9402541630 0.888148824  1.22712132
#>    56   90 0.7612457790 0.991453475 0.7786688131 1.000000000  0.58204591
#>    90   63 0.7786688131 1.000000000 0.4262983784 0.967513254  0.55386901
#>    55   56 0.7787101765 0.983351877 0.7612457790 0.991453475  0.75148055
#>    21   55 0.8015995365 0.879146532 0.7787101765 0.983351877  0.76132574
#>    90   66 0.7786688131 1.000000000 0.1575476249 0.960627633  0.17170607
#>    55   76 0.7787101765 0.983351877 0.9050062503 0.895773977  0.87816650
#>    90   76 0.7786688131 1.000000000 0.9050062503 0.895773977  0.96774419
#>    85   60 1.0000000000 0.367663354 0.9402541630 0.888148824  2.10429477
#>    73   22 0.5150122847 0.388362764 0.3939444383 0.415222653  0.46038162
#>    50    2 0.5369571944 0.465667446 0.5278121866 0.422022382  0.45881414
#>    62   18 0.3446781846 0.258733578 0.4100521792 0.372344619  0.38150523
#>    18   30 0.4100521792 0.372344619 0.2763887113 0.422632833  0.33192219
#>    89   30 0.2705497243 0.630117475 0.2763887113 0.422632833  0.33790894
#>    73    2 0.5150122847 0.388362764 0.5278121866 0.422022382  0.60204892
#>    25    1 0.5320513438 0.179344111 0.4466615437 0.321687757  0.54829468
#>    22   15 0.3939444383 0.415222653 0.4678868140 0.592247867  0.44523643
#>    18   38 0.4100521792 0.372344619 0.2510969860 0.384233831  0.32922909
#>    73   57 0.5150122847 0.388362764 0.6848346683 0.404749044  0.60215096
#>    30   41 0.2763887113 0.422632833 0.2587567555 0.429048097  0.27925468
#>    99   50 0.5702165733 0.681418154 0.5369571944 0.465667446  0.57740200
#>    71   73 0.6505488198 0.250454038 0.5150122847 0.388362764  0.56125767
#>     2   87 0.5278121866 0.422022382 0.6521022497 0.478622504  0.60641182
#>    57   26 0.6848346683 0.404749044 0.8111563873 0.452032598  0.76902730
#>    54   41 0.1929083827 0.401929694 0.2587567555 0.429048097  0.22754928
#>    22   89 0.3939444383 0.415222653 0.2705497243 0.630117475  0.35470738
#>    77   54 0.2450172421 0.448829346 0.1929083827 0.401929694  0.21249858
#>    67   97 0.7070090079 0.214810537 0.6395240227 0.188045164  0.66779256
#>    16   62 0.4069079741 0.006799476 0.3446781846 0.258733578  0.40252441
#>    80   64 0.7095124588 0.450016374 0.7403136379 0.608850651  0.75890986
#>    80   57 0.7095124588 0.450016374 0.6848346683 0.404749044  0.66983018
#>    38   48 0.2510969860 0.384233831 0.1600454796 0.337072368  0.20971047
#>     5    9 0.5815826525 0.116122553 0.5503721815 0.131553705  0.58866220
#>    67   91 0.7070090079 0.214810537 0.6620263304 0.308433610  0.73773322
#>    23   69 0.1781907247 0.484514858 0.2015611267 0.500415584  0.20511918
#>     7   25 0.5188617954 0.075654331 0.5320513438 0.179344111  0.48254329
#>    33   58 0.0850819863 0.147099916 0.2711813997 0.098344998  0.19576289
#>    38   36 0.2510969860 0.384233831 0.3160308839 0.244458235  0.24003161
#>    49   54 0.0496714908 0.423565645 0.1929083827 0.401929694  0.11914547
#>    23    6 0.1781907247 0.484514858 0.2344680082 0.461534878  0.19483568
#>    97  100 0.6395240227 0.188045164 0.6301912396 0.105026970  0.61461506
#>   100    7 0.6301912396 0.105026970 0.5188617954 0.075654331  0.58886389
#>    69   20 0.2015611267 0.500415584 0.2171869312 0.619036821  0.25868323
#>    89   77 0.2705497243 0.630117475 0.2450172421 0.448829346  0.32915653
#>    12   11 0.2622078278 0.817709859 0.3199489062 0.674808412  0.23199195
#>    23   49 0.1781907247 0.484514858 0.0496714908 0.423565645  0.07848020
#>    69   53 0.2015611267 0.500415584 0.1764550764 0.576789613  0.24071012
#>    80   87 0.7095124588 0.450016374 0.6521022497 0.478622504  0.71427960
#>    67   71 0.7070090079 0.214810537 0.6505488198 0.250454038  0.70254974
#>    67   51 0.7070090079 0.214810537 0.8159650173 0.295102158  0.76351823
#>     7   96 0.5188617954 0.075654331 0.6682576706 0.021180060  0.54327644
#>    58   62 0.2711813997 0.098344998 0.3446781846 0.258733578  0.39558271
#>    81   15 0.5357936883 0.766838802 0.4678868140 0.592247867  0.43426573
#>    19   17 0.6573278026 0.771914724 0.6839011949 0.726409285  0.62068958
#>    61   92 0.9586015217 0.514597579 0.9532668571 0.428360448  0.88812130
#>    46   75 0.8729073784 0.320999731 0.8639968879 0.304961370  0.83331615
#>    67   29 0.7070090079 0.214810537 0.8312221603 0.249472198  0.78287143
#>    58   44 0.2711813997 0.098344998 0.2590521812 0.014244331  0.23549459
#>    33   36 0.0850819863 0.147099916 0.3160308839 0.244458235  0.19853040
#>    46   51 0.8729073784 0.320999731 0.8159650173 0.295102158  0.82744947
#>    58   95 0.2711813997 0.098344998 0.2770127215 0.000000000  0.30272859
#>    58    4 0.2711813997 0.098344998 0.3541497732 0.014432560  0.30736207
#>    84   44 0.1093853138 0.069653514 0.2590521812 0.014244331  0.08666026
#>    78   52 0.7851591563 0.746433158 0.8080388551 0.623736844  0.76053395
#>    14   84 0.1847591012 0.071820591 0.1093853138 0.069653514  0.14555406
#>    19   99 0.6573278026 0.771914724 0.5702165733 0.681418154  0.61000122
#>   100   47 0.6301912396 0.105026970 0.8480127753 0.101363319  0.73911409
#>    37   27 0.4485215172 0.845474257 0.5333868682 0.821445378  0.47912959
#>    12   37 0.2622078278 0.817709859 0.4485215172 0.845474257  0.36370771
#>    32   94 0.5022711789 0.862439104 0.6035318922 0.820730584  0.56950931
#>    94   74 0.6035318922 0.820730584 0.5788572291 0.785086372  0.57254716
#>    67   34 0.7070090079 0.214810537 0.8635004214 0.166238918  0.77372699
#>     3   47 0.8881957112 0.167826981 0.8480127753 0.101363319  0.87833165
#>    37   81 0.4485215172 0.845474257 0.5357936883 0.766838802  0.42382351
#>     8   33 0.0004357169 0.272443903 0.0850819863 0.147099916 -0.28501078
#>     8   48 0.0004357169 0.272443903 0.1600454796 0.337072368  0.10239487
#>    10   68 0.9555826481 0.382968668 0.9383035630 0.287289448  0.92373753
#>    75   68 0.8639968879 0.304961370 0.9383035630 0.287289448  0.89675308
#>    28   29 0.9206979757 0.242333838 0.8312221603 0.249472198  0.87707004
#>    92   10 0.9532668571 0.428360448 0.9555826481 0.382968668  0.90987413
#>    20   40 0.2171869312 0.619036821 0.1040421188 0.741456118  0.20931572
#>    12   89 0.2622078278 0.817709859 0.2705497243 0.630117475  0.22363602
#>    45   63 0.3053016205 0.883162585 0.4262983784 0.967513254  0.38552366
#>    32   86 0.5022711789 0.862439104 0.5412423298 0.817867353  0.56312574
#>    10   46 0.9555826481 0.382968668 0.8729073784 0.320999731  0.91128831
#>    42   19 0.7067057662 0.952000767 0.6573278026 0.771914724  0.68840056
#>    19   93 0.6573278026 0.771914724 0.7936747403 0.828611652  0.72815971
#>    78   64 0.7851591563 0.746433158 0.7403136379 0.608850651  0.74501133
#>    61   26 0.9586015217 0.514597579 0.8111563873 0.452032598  0.86132505
#>    68   85 0.9383035630 0.287289448 1.0000000000 0.367663354  1.06296948
#>    59   28 0.9445794478 0.205076860 0.9206979757 0.242333838  0.97125660
#>    59    3 0.9445794478 0.205076860 0.8881957112 0.167826981  0.90393243
#>    52   83 0.8080388551 0.623736844 0.8157843325 0.726831319  0.84052604
#>    52   88 0.8080388551 0.623736844 0.8972254874 0.656380294  0.88016996
#>    12   43 0.2622078278 0.817709859 0.2345564400 0.869332644  0.27367791
#>    92   85 0.9532668571 0.428360448 1.0000000000 0.367663354  0.98885407
#>     3   79 0.8881957112 0.167826981 0.9465931931 0.102012070  0.89720220
#>    35   83 0.8462844438 0.827614859 0.8157843325 0.726831319  0.82719573
#>    47   13 0.8480127753 0.101363319 0.9054318963 0.017935214  0.78654416
#>    13   16 0.9054318963 0.017935214 0.4069079741 0.006799476  0.73714848
#>    13   95 0.9054318963 0.017935214 0.2770127215 0.000000000  0.89040814
#>    98   45 0.2986739179 0.918012248 0.3053016205 0.883162585  0.26649057
#>    63   70 0.4262983784 0.967513254 0.5111976552 0.949869835  0.45884605
#>    42   94 0.7067057662 0.952000767 0.6035318922 0.820730584  0.63156418
#>    35   78 0.8462844438 0.827614859 0.7851591563 0.746433158  0.81915214
#>    60   88 0.9402541630 0.888148824 0.8972254874 0.656380294  0.95245492
#>    63   39 0.4262983784 0.967513254 0.5568987646 0.952751682  0.51848949
#>    31   21 0.7969770907 0.850551389 0.8015995365 0.879146532  0.81358309
#>    31   65 0.7969770907 0.850551389 0.8340500675 0.863276476  0.82163257
#>    72   24 0.0000000000 0.794998361 0.0169148210 0.403047325 -1.53436718
#>    72   49 0.0000000000 0.794998361 0.0496714908 0.423565645 -0.08062880
#>    66   40 0.1575476249 0.960627633 0.1040421188 0.741456118  0.09867455
#>    66   43 0.1575476249 0.960627633 0.2345564400 0.869332644  0.12261867
#>    39   56 0.5568987646 0.952751682 0.7612457790 0.991453475  0.63260173
#>    19   31 0.6573278026 0.771914724 0.7969770907 0.850551389  0.70339780
#>    66   63 0.1575476249 0.960627633 0.4262983784 0.967513254  0.28751673
#>    60   35 0.9402541630 0.888148824 0.8462844438 0.827614859  0.90973000
#>    39   90 0.5568987646 0.952751682 0.7786688131 1.000000000  0.58204591
#>    42   55 0.7067057662 0.952000767 0.7787101765 0.983351877  0.76132574
#>    90   55 0.7786688131 1.000000000 0.7787101765 0.983351877  0.77195255
#>    60   76 0.9402541630 0.888148824 0.9050062503 0.895773977  0.96774419
#>    50   22 0.5369571944 0.465667446 0.3939444383 0.415222653  0.45881414
#>    57   51 0.6848346683 0.404749044 0.8159650173 0.295102158  0.76902730
#>    54   38 0.1929083827 0.401929694 0.2510969860 0.384233831  0.22754928
#>    16    7 0.4069079741 0.006799476 0.5188617954 0.075654331  0.40252441
#>     5   97 0.5815826525 0.116122553 0.6395240227 0.188045164  0.58866220
#>     9    7 0.5503721815 0.131553705 0.5188617954 0.075654331  0.55131844
#>     6   89 0.2344680082 0.461534878 0.2705497243 0.630117475  0.28613488
#>    33   14 0.0850819863 0.147099916 0.1847591012 0.071820591  0.19576289
#>    64   99 0.7403136379 0.608850651 0.5702165733 0.681418154  0.63088365
#>    84   95 0.1093853138 0.069653514 0.2770127215 0.000000000  0.08666026
#>    17   78 0.6839011949 0.726409285 0.7851591563 0.746433158  0.72565469
#>    81   86 0.5357936883 0.766838802 0.5412423298 0.817867353  0.52551102
#>    27   32 0.5333868682 0.821445378 0.5022711789 0.862439104  0.48665650
#>     8   84 0.0004357169 0.272443903 0.1093853138 0.069653514 -0.28501078
#>    75   28 0.8639968879 0.304961370 0.9206979757 0.242333838  0.89675308
#>    20   12 0.2171869312 0.619036821 0.2622078278 0.817709859  0.20931572
#>    45   37 0.3053016205 0.883162585 0.4485215172 0.845474257  0.38552366
#>    61   52 0.9586015217 0.514597579 0.8080388551 0.623736844  0.86132505
#>    68   59 0.9383035630 0.287289448 0.9445794478 0.205076860  1.06296948
#>    79   85 0.9465931931 0.102012070 1.0000000000 0.367663354  1.33938577
#>    98   43 0.2986739179 0.918012248 0.2345564400 0.869332644  0.26649057
#>    42   39 0.7067057662 0.952000767 0.5568987646 0.952751682  0.63156418
#>    93   35 0.7936747403 0.828611652 0.8462844438 0.827614859  0.82012615
#>    72    8 0.0000000000 0.794998361 0.0004357169 0.272443903 -1.53436718
#>    66   72 0.1575476249 0.960627633 0.0000000000 0.794998361  0.09867455
#>    60   90 0.9402541630 0.888148824 0.7786688131 1.000000000  0.96774419
#>            my1         mx2           my2 bp1 bp2
#>    0.372572672  0.45731152   0.414562826   0   0
#>    0.428400973  0.45881414   0.459260247   0   0
#>    0.414562826  0.33192219   0.367458550   0   0
#>    0.313156836  0.46371971   0.372572672   0   0
#>    0.343941606  0.29616759   0.320201166   0   0
#>    0.367458550  0.34206125   0.528305456   0   0
#>    0.528188603  0.27925468   0.457947560   0   0
#>    0.320201166  0.38416866   0.143603435   0   0
#>    0.374528481  0.60641182   0.414189243   0   0
#>    0.285904322  0.54829468   0.285872090   0   0
#>    0.285872090  0.46371971   0.372572672   0   0
#>    0.220229399  0.38150523   0.313156836   0   0
#>    0.142692186  0.43886920   0.220229399   0   0
#>    0.429825932  0.58364267   0.568909505   0   0
#>    0.405730142  0.22754928   0.411320412   0   0
#>    0.497753546  0.57740200   0.569871544   0   0
#>    0.360300438  0.26025573   0.405730142   0   0
#>    0.373470946  0.74233050   0.340275204   0   0
#>    0.457947560  0.22411833   0.419651368   0   0
#>    0.298255739  0.60215096   0.373470946   0   0
#>    0.414189243  0.66983018   0.442289137   0   0
#>    0.372202795  0.78070902   0.372560744   0   0
#>    0.535566989  0.36280925   0.577785491   0   0
#>    0.432561744  0.19483568   0.444877174   0   0
#>    0.215229631  0.58200769   0.230383898   0   0
#>    0.230383898  0.54869008   0.285904322   0   0
#>    0.172873307  0.58200769   0.230383898   0   0
#>    0.139369384  0.40601691   0.142692186   0   0
#>    0.522840843  0.79148306   0.537555826   0   0
#>    0.352661754  0.11914547   0.398550614   0   0
#>    0.169719518  0.58666378   0.172873307   0   0
#>    0.094189435  0.58886389   0.035998529   0   0
#>    0.287190323  0.74233050   0.340275204   0   0
#>    0.139455042  0.73922489   0.110501184   0   0
#>    0.470061153  0.28613488   0.538629204   0   0
#>    0.538629204  0.25868323   0.553230773   0   0
#>    0.132957876  0.58666378   0.172873307   0   0
#>    0.190021406  0.23549459   0.060566856   0   0
#>    0.294122736  0.19270279   0.214409137   0   0
#>    0.148811707  0.58886389   0.035998529   0   0
#>    0.553230773  0.22363602   0.722012974   0   0
#>    0.529421286  0.21249858   0.432561744   0   0
#>    0.691314338  0.57740200   0.569871544   0   0
#>    0.577785491  0.23199195   0.722384547   0   0
#>    0.138506774  0.35192982  -0.187054322   0   0
#>    0.530402841  0.07848020   0.528793047   0   0
#>    0.528793047 -0.00293336   0.596157850   0   0
#>    0.555598344  0.13550005   0.657034667   0   0
#>    0.531495553  0.63088365   0.587984666   0   0
#>    0.392315269  0.75890986   0.522840843   0   0
#>    0.270285866  0.56125767   0.298255739   0   0
#>    0.252199974  0.84425767   0.279196560   0   0
#>    0.587984666  0.58364267   0.568909505   0   0
#>   -0.089484899  0.73714848  -3.612874244   0   0
#>    0.138373038  0.39903218   0.138506774   0   0
#>    0.110501184  0.77372699   0.153383912   0   0
#>    0.705826342  0.36280925   0.577785491   0   0
#>    0.720007812  0.65270027   0.639122473   0   0
#>    0.475673947  0.88396343   0.450713090   0   0
#>    0.372560744  0.88396343   0.450713090   0   0
#>    0.332501175  0.84425767   0.279196560   0   0
#>    0.182846083  0.87259366   0.217640791   0   0
#>    0.035731813  0.73911409   0.103913337   0   0
#>    0.060566856  0.15051716  -0.049083105   0   0
#>    0.200585131  0.38416866   0.143603435   0   0
#>    0.345400553  0.78070902   0.372560744   0   0
#>    0.050870192  0.08666026  -0.221568291   0   0
#>    0.051144933  0.35192982  -0.187054322   0   0
#>   -0.187054322  0.89040814 -10.473991484   0   0
#>    0.639122473  0.74501133   0.683419454   0   0
#>    0.678359799  0.79148306   0.537555826   0   0
#>    0.344218609  0.11914547   0.398550614   0   0
#>    0.432296687  0.19483568   0.444877174   0   0
#>    0.733175673  0.47157659   0.691314338   0   0
#>    0.123540065 -0.28501078  -0.011574777   0   0
#>    0.730296343  0.57545581   0.733175673   0   0
#>    0.781303319  0.72815971   0.793870065   0   0
#>    0.330005658  0.07656883   0.344218609   0   0
#>    0.601439129 -0.00293336   0.596157850   0   0
#>    0.218462255  0.87259366   0.217640791   0   0
#>    0.217640791  0.87833165   0.128411813   0   0
#>    0.103913337  0.78654416  -0.002415568   0   0
#>    0.791697625  0.52551102   0.793741916   0   0
#>    0.793741916  0.56312574   0.876324050   0   0
#>    0.791026451  0.57545581   0.733175673   0   0
#>    0.783685880  0.61000122   0.730296343   0   0
#>    0.775605909  0.42382351   0.730317143   0   0
#>    0.881905504  0.57254716   0.815817120   0   0
#>    0.815817120  0.55094236   0.791026451   0   0
#>    0.153383912  0.87833165   0.128411813   0   0
#>    0.294967364  0.92373753   0.339319843   0   0
#>    0.730317143  0.43426573   0.705826342   0   0
#>    0.818281190  0.45765087   0.910179601   0   0
#>    0.250044814  0.19270279   0.214409137   0   0
#>    0.277636345  0.97125660   0.248459157   0   0
#>    0.259815908  0.84425767   0.279196560   0   0
#>    0.403391681  0.91128831   0.355928860   0   0
#>    0.657034667 -0.00293336   0.596157850   0   0
#>    0.725258066  0.17201944   0.802618105   0   0
#>    0.897045331  0.45765087   0.910179601   0   0
#>    0.355928860  0.82744947   0.345400553   0   0
#>    0.450713090  0.90987413   0.403391681   0   0
#>    0.860207371  0.61896290   0.783685880   0   0
#>    0.793870065  0.81915214   0.784441155   0   0
#>    0.678918841  0.82719573   0.778384783   0   0
#>    0.538823906  0.79148306   0.537555826   0   0
#>    0.255460268  1.33938577   0.161238768   0   0
#>    0.205304814  0.87377085   0.218462255   0   0
#>    0.153400594  1.33938577   0.161238768   0   0
#>    0.673134281  0.90991758   0.753350678   0   0
#>    0.464830082  2.10429477   0.758095694   0   0
#>    0.564821332  1.22712132   0.715012406   0   0
#>    0.857070757  0.26649057   0.893836566   0   0
#>    0.800209474  0.27367791   0.857070757   0   0
#>    0.407421068  0.96294777   0.332238714   0   0
#>    0.802618105  0.12261867   0.853037900   0   0
#>    0.117002963  0.89748559   0.073939511   0   0
#>    0.073939511 10.49182821  -4.623138895   0   1
#>    0.883138958  0.53709114   0.903055118   0   0
#>   -0.002415568  0.73714848  -3.612874244   0   0
#>    0.914199898  0.28751673   1.136050366   0   0
#>    0.911043761  0.53709114   0.903055118   0   0
#>    0.904878822  0.57001736   0.883138958   0   0
#>    0.903055118  0.51848949   1.198045120   0   0
#>    0.835848610  0.82163257   0.839087002   0   0
#>    0.856584018  0.72449180   0.876939925   0   0
#>    0.766005229  0.90991758   0.753350678   0   0
#>    0.910179601  0.45884605   0.911043761   0   0
#>    0.876939925  0.76132574   0.924916707   0   0
#>    0.862538190  0.84649058   0.929826280   0   0
#>    0.839087002  0.87903013   0.858778294   0   0
#>    0.858778294  0.84649058   0.929826280   0   0
#>    0.532441563  0.07001749   0.330005658   0   0
#>    0.595178288  0.07656883   0.344218609   0   0
#>    0.858883287 -0.03381158   0.601439129   0   0
#>    1.111868112  0.75148055   0.947528261   0   0
#>    0.853418462  0.82012615   0.835848610   0   0
#>    0.943153001  0.84649058   0.929826280   0   0
#>    0.942945752  0.28751673   1.136050366   0   0
#>    0.832329124  0.87903013   0.858778294   0   0
#>    0.715012406  2.10429477   0.758095694   0   0
#>    1.378804843  0.77195255   0.991659200   0   0
#>    1.511059136  0.17170607   5.656224073   0   0
#>    0.947528261  0.77195255   0.991659200   0   0
#>    0.924916707  0.84434769   0.943153001   0   0
#>    5.656224073  0.07005844   7.259772481   0   1
#>    0.991923096  0.84434769   0.943153001   0   0
#>    1.100504578  0.87816650   0.991923096   0   0
#>    0.758095694  4.00059155   0.975769087   0   1
#>    0.428400973  0.45731152   0.414562826   0   0
#>    0.459260247  0.59929107   0.429825932   0   0
#>    0.313156836  0.32800552   0.343941606   0   0
#>    0.367458550  0.32922909   0.360300438   0   0
#>    0.528188603  0.34206125   0.528305456   0   0
#>    0.374528481  0.46038162   0.428400973   0   0
#>    0.285872090  0.43886920   0.220229399   0   0
#>    0.497753546  0.35470738   0.535566989   0   0
#>    0.360300438  0.32800552   0.343941606   0   0
#>    0.373470946  0.60204892   0.374528481   0   0
#>    0.457947560  0.26025573   0.405730142   0   0
#>    0.569871544  0.58364267   0.568909505   0   0
#>    0.298255739  0.54869008   0.285904322   0   0
#>    0.414189243  0.59929107   0.429825932   0   0
#>    0.372202795  0.76149899   0.392315269   0   0
#>    0.411320412  0.22411833   0.419651368   0   0
#>    0.535566989  0.34206125   0.528305456   0   0
#>    0.432561744  0.22411833   0.419651368   0   0
#>    0.215229631  0.69784569   0.139455042   0   0
#>    0.139369384  0.39903218   0.138506774   0   0
#>    0.522840843  0.71427960   0.531495553   0   0
#>    0.442289137  0.76149899   0.392315269   0   0
#>    0.352661754  0.24003161   0.294122736   0   0
#>    0.169719518  0.55131844   0.094189435   0   0
#>    0.287190323  0.70254974   0.270285866   0   0
#>    0.470061153  0.16406397   0.530402841   0   0
#>    0.132957876  0.40601691   0.142692186   0   0
#>    0.190021406  0.19853040   0.200585131   0   0
#>    0.294122736  0.29616759   0.320201166   0   0
#>    0.398550614  0.12424282   0.432296687   0   0
#>    0.444877174  0.20511918   0.470061153   0   0
#>    0.148811707  0.69784569   0.139455042   0   0
#>    0.035998529  0.58893426   0.035731813   0   0
#>    0.553230773  0.24071012   0.555598344   0   0
#>    0.529421286  0.33790894   0.528188603   0   0
#>    0.722384547  0.36370771   0.775605909   0   0
#>    0.528793047  0.12424282   0.432296687   0   0
#>    0.555598344  0.16406397   0.530402841   0   0
#>    0.531495553  0.66983018   0.442289137   0   0
#>    0.270285866  0.66779256   0.215229631   0   0
#>    0.252199974  0.73773322   0.287190323   0   0
#>   -0.089484899  0.58893426   0.035731813   0   0
#>    0.138373038  0.38416866   0.143603435   0   0
#>    0.705826342  0.47157659   0.691314338   0   0
#>    0.720007812  0.72565469   0.781303319   0   0
#>    0.475673947  1.06341701   0.464830082   0   0
#>    0.332501175  0.90087481   0.294967364   0   0
#>    0.182846083  0.76351823   0.252199974   0   0
#>    0.060566856  0.30272859   0.050870192   0   0
#>    0.200585131  0.19270279   0.214409137   0   0
#>    0.345400553  0.83331615   0.332501175   0   0
#>    0.050870192  0.30736207   0.051144933   0   0
#>    0.051144933  0.39558271   0.138373038   0   0
#>   -0.221568291  0.15051716  -0.049083105   0   0
#>    0.678359799  0.76353192   0.678918841   0   0
#>    0.123540065  0.15051716  -0.049083105   0   0
#>    0.730296343  0.62068958   0.720007812   0   0
#>    0.103913337  0.73922489   0.110501184   0   0
#>    0.791697625  0.48665650   0.818281190   0   0
#>    0.775605909  0.36004129   0.800209474   0   0
#>    0.881905504  0.57001736   0.883138958   0   0
#>    0.815817120  0.61896290   0.783685880   0   0
#>    0.153383912  0.78287143   0.182846083   0   0
#>    0.128411813  0.89720220   0.117002963   0   0
#>    0.730317143  0.47912959   0.791697625   0   0
#>   -0.011574777  0.10239487   0.250044814   0   0
#>    0.250044814  0.07001749   0.330005658   0   0
#>    0.339319843  0.96294777   0.332238714   0   0
#>    0.277636345  0.90087481   0.294967364   0   0
#>    0.259815908  0.87377085   0.218462255   0   0
#>    0.403391681  0.98885407   0.407421068   0   0
#>    0.725258066  0.13550005   0.657034667   0   0
#>    0.722012974  0.23199195   0.722384547   0   0
#>    0.897045331  0.37356467   0.914199898   0   0
#>    0.876324050  0.56950931   0.881905504   0   0
#>    0.355928860  0.92373753   0.339319843   0   0
#>    0.860207371  0.70161527   0.856584018   0   0
#>    0.793870065  0.70339780   0.853418462   0   0
#>    0.683419454  0.76053395   0.678359799   0   0
#>    0.538823906  0.88812130   0.475673947   0   0
#>    0.255460268  0.96294777   0.332238714   0   0
#>    0.248459157  0.90393243   0.205304814   0   0
#>    0.205304814  0.93822292   0.153400594   0   0
#>    0.673134281  0.76353192   0.678918841   0   0
#>    0.564821332  0.84052604   0.673134281   0   0
#>    0.857070757  0.17201944   0.802618105   0   0
#>    0.407421068  1.06341701   0.464830082   0   0
#>    0.117002963  0.93822292   0.153400594   0   0
#>    0.778384783  0.90991758   0.753350678   0   0
#>   -0.002415568  0.89748559   0.073939511   0   0
#>   -3.612874244  0.89040814 -10.473991484   0   0
#>  -10.473991484  0.93578706 -12.063990897   0   1
#>    0.893836566  0.37356467   0.914199898   0   0
#>    0.911043761  0.51848949   1.198045120   0   0
#>    0.904878822  0.68840056   0.860207371   0   0
#>    0.784441155  0.82719573   0.778384783   0   0
#>    0.766005229  1.22712132   0.715012406   0   0
#>    1.198045120  0.55386901   1.511059136   0   0
#>    0.862538190  0.72449180   0.876939925   0   0
#>    0.839087002  0.81358309   0.862538190   0   0
#>    0.532441563 -0.08062880   0.595178288   0   0
#>    0.595178288 -0.03381158   0.601439129   0   0
#>    0.858883287  0.12261867   0.853037900   0   0
#>    0.853037900  0.22920564   0.942945752   0   0
#>    1.111868112  0.58204591   1.378804843   0   0
#>    0.853418462  0.70161527   0.856584018   0   0
#>    1.136050366  0.17170607   5.656224073   0   0
#>    0.832329124  0.95245492   0.766005229   0   0
#>    1.378804843  0.55386901   1.511059136   0   0
#>    0.924916707  0.75148055   0.947528261   0   0
#>    0.991659200  0.87816650   0.991923096   0   0
#>    1.100504578  0.90973000   0.832329124   0   0
#>    0.459260247  0.44523643   0.497753546   0   0
#>    0.372202795  0.74233050   0.340275204   0   0
#>    0.411320412  0.20971047   0.352661754   0   0
#>    0.139369384  0.54327644  -0.089484899   0   0
#>    0.169719518  0.61461506   0.148811707   0   0
#>    0.094189435  0.48254329   0.132957876   0   0
#>    0.538629204  0.32915653   0.529421286   0   0
#>    0.190021406  0.14555406   0.123540065   0   0
#>    0.587984666  0.65270027   0.639122473   0   0
#>   -0.221568291 -2.02721929  -5.308808334   0   1
#>    0.781303319  0.74501133   0.683419454   0   0
#>    0.793741916  0.55094236   0.791026451   0   0
#>    0.818281190  0.56312574   0.876324050   0   0
#>   -0.011574777 -4.11168088  -2.067461986   0   1
#>    0.277636345  0.87707004   0.259815908   0   0
#>    0.725258066  0.22363602   0.722012974   0   0
#>    0.897045331  0.36004129   0.800209474   0   0
#>    0.538823906  0.88016996   0.564821332   0   0
#>    0.255460268  0.97125660   0.248459157   0   0
#>    0.161238768  4.95748467  -0.566147550   0   1
#>    0.893836566  0.22920564   0.942945752   0   0
#>    0.904878822  0.63260173   1.111868112   0   0
#>    0.835848610  0.81915214   0.784441155   0   0
#>    0.532441563 -3.44804200   0.530845900   0   1
#>    0.858883287 -3.07100222   3.873900151   0   1
#>    1.100504578  3.86390412   5.284430320   0   1
#> 
#> $x
#>                [,1]        [,2]
#>   [1,] 0.4466615437 0.321687757
#>   [2,] 0.5278121866 0.422022382
#>   [3,] 0.8881957112 0.167826981
#>   [4,] 0.3541497732 0.014432560
#>   [5,] 0.5815826525 0.116122553
#>   [6,] 0.2344680082 0.461534878
#>   [7,] 0.5188617954 0.075654331
#>   [8,] 0.0004357169 0.272443903
#>   [9,] 0.5503721815 0.131553705
#>  [10,] 0.9555826481 0.382968668
#>  [11,] 0.3199489062 0.674808412
#>  [12,] 0.2622078278 0.817709859
#>  [13,] 0.9054318963 0.017935214
#>  [14,] 0.1847591012 0.071820591
#>  [15,] 0.4678868140 0.592247867
#>  [16,] 0.4069079741 0.006799476
#>  [17,] 0.6839011949 0.726409285
#>  [18,] 0.4100521792 0.372344619
#>  [19,] 0.6573278026 0.771914724
#>  [20,] 0.2171869312 0.619036821
#>  [21,] 0.8015995365 0.879146532
#>  [22,] 0.3939444383 0.415222653
#>  [23,] 0.1781907247 0.484514858
#>  [24,] 0.0169148210 0.403047325
#>  [25,] 0.5320513438 0.179344111
#>  [26,] 0.8111563873 0.452032598
#>  [27,] 0.5333868682 0.821445378
#>  [28,] 0.9206979757 0.242333838
#>  [29,] 0.8312221603 0.249472198
#>  [30,] 0.2763887113 0.422632833
#>  [31,] 0.7969770907 0.850551389
#>  [32,] 0.5022711789 0.862439104
#>  [33,] 0.0850819863 0.147099916
#>  [34,] 0.8635004214 0.166238918
#>  [35,] 0.8462844438 0.827614859
#>  [36,] 0.3160308839 0.244458235
#>  [37,] 0.4485215172 0.845474257
#>  [38,] 0.2510969860 0.384233831
#>  [39,] 0.5568987646 0.952751682
#>  [40,] 0.1040421188 0.741456118
#>  [41,] 0.2587567555 0.429048097
#>  [42,] 0.7067057662 0.952000767
#>  [43,] 0.2345564400 0.869332644
#>  [44,] 0.2590521812 0.014244331
#>  [45,] 0.3053016205 0.883162585
#>  [46,] 0.8729073784 0.320999731
#>  [47,] 0.8480127753 0.101363319
#>  [48,] 0.1600454796 0.337072368
#>  [49,] 0.0496714908 0.423565645
#>  [50,] 0.5369571944 0.465667446
#>  [51,] 0.8159650173 0.295102158
#>  [52,] 0.8080388551 0.623736844
#>  [53,] 0.1764550764 0.576789613
#>  [54,] 0.1929083827 0.401929694
#>  [55,] 0.7787101765 0.983351877
#>  [56,] 0.7612457790 0.991453475
#>  [57,] 0.6848346683 0.404749044
#>  [58,] 0.2711813997 0.098344998
#>  [59,] 0.9445794478 0.205076860
#>  [60,] 0.9402541630 0.888148824
#>  [61,] 0.9586015217 0.514597579
#>  [62,] 0.3446781846 0.258733578
#>  [63,] 0.4262983784 0.967513254
#>  [64,] 0.7403136379 0.608850651
#>  [65,] 0.8340500675 0.863276476
#>  [66,] 0.1575476249 0.960627633
#>  [67,] 0.7070090079 0.214810537
#>  [68,] 0.9383035630 0.287289448
#>  [69,] 0.2015611267 0.500415584
#>  [70,] 0.5111976552 0.949869835
#>  [71,] 0.6505488198 0.250454038
#>  [72,] 0.0000000000 0.794998361
#>  [73,] 0.5150122847 0.388362764
#>  [74,] 0.5788572291 0.785086372
#>  [75,] 0.8639968879 0.304961370
#>  [76,] 0.9050062503 0.895773977
#>  [77,] 0.2450172421 0.448829346
#>  [78,] 0.7851591563 0.746433158
#>  [79,] 0.9465931931 0.102012070
#>  [80,] 0.7095124588 0.450016374
#>  [81,] 0.5357936883 0.766838802
#>  [82,] 0.8654626457 0.380283500
#>  [83,] 0.8157843325 0.726831319
#>  [84,] 0.1093853138 0.069653514
#>  [85,] 1.0000000000 0.367663354
#>  [86,] 0.5412423298 0.817867353
#>  [87,] 0.6521022497 0.478622504
#>  [88,] 0.8972254874 0.656380294
#>  [89,] 0.2705497243 0.630117475
#>  [90,] 0.7786688131 1.000000000
#>  [91,] 0.6620263304 0.308433610
#>  [92,] 0.9532668571 0.428360448
#>  [93,] 0.7936747403 0.828611652
#>  [94,] 0.6035318922 0.820730584
#>  [95,] 0.2770127215 0.000000000
#>  [96,] 0.6682576706 0.021180060
#>  [97,] 0.6395240227 0.188045164
#>  [98,] 0.2986739179 0.918012248
#>  [99,] 0.5702165733 0.681418154
#> [100,] 0.6301912396 0.105026970
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,73,18), [3,5,12]
#> 2: (73,2,22), [4,3,10]
#> 3: (73,22,18), [7,1,2]
#> 4: (50,22,2), [2,15,17]
#> 5: (62,1,18), [1,6,13]
#> 6: (18,38,62), [9,5,18]
#> 7: (18,22,30), [28,18,3]
#> 8: (89,77,30), [20,28,52]
#> 9: (38,36,62), [67,6,46]
#> 10: (73,57,2), [23,2,19]
#> 11: (71,73,25), [12,30,22]
#> 12: (25,73,1), [1,13,11]
#> 13: (25,1,62), [5,14,12]
#> 14: (7,25,62), [13,32,44]
#> 15: (2,87,50), [54,4,23]
#> 16: (30,41,38), [25,18,20]
#> 17: (22,50,15), [21,26,4]
#> 18: (18,30,38), [16,6,7]
#> 19: (73,91,57), [35,10,22]
#> 20: (30,77,41), [37,16,8]
#> 21: (99,15,50), [17,54,53]
#> 22: (71,91,73), [19,11,63]
#> 23: (2,57,87), [34,15,10]
#> 24: (57,51,26), [74,62,35]
#> 25: (54,38,41), [16,37,36]
#> 26: (22,15,89), [55,28,17]
#> 27: (77,6,54), [48,37,52]
#> 28: (89,30,22), [7,26,8]
#> 29: (67,71,97), [30,41,63]
#> 30: (97,71,25), [11,31,29]
#> 31: (9,97,25), [30,44,38]
#> 32: (16,7,62), [14,56,66]
#> 33: (80,26,64), [89,61,62]
#> 34: (80,87,57), [23,62,61]
#> 35: (51,57,91), [19,40,24]
#> 36: (38,54,48), [47,46,25]
#> 37: (54,41,77), [20,27,25]
#> 38: (5,97,9), [31,39,49]
#> 39: (9,7,5), [50,38,44]
#> 40: (67,51,91), [35,63,64]
#> 41: (97,100,67), [70,29,49]
#> 42: (23,6,69), [43,57,48]
#> 43: (6,89,69), [51,42,52]
#> 44: (7,9,25), [31,14,39]
#> 45: (33,14,58), [78,79,91]
#> 46: (38,48,36), [92,9,36]
#> 47: (49,48,54), [36,88,87]
#> 48: (23,54,6), [27,42,88]
#> 49: (97,5,100), [50,41,38]
#> 50: (100,5,7), [39,77,49]
#> 51: (69,89,20), [121,60,43]
#> 52: (89,6,77), [27,8,43]
#> 53: (81,15,99), [21,90,71]
#> 54: (99,50,87), [15,65,21]
#> 55: (15,11,89), [58,26,71]
#> 56: (62,4,16), [83,32,68]
#> 57: (69,53,23), [59,42,60]
#> 58: (12,89,11), [55,104,121]
#> 59: (23,53,49), [137,88,57]
#> 60: (69,20,53), [119,57,51]
#> 61: (80,64,87), [65,34,33]
#> 62: (57,26,80), [33,34,24]
#> 63: (67,91,71), [22,29,40]
#> 64: (67,29,51), [69,40,76]
#> 65: (64,99,87), [54,61,85]
#> 66: (7,16,96), [151,77,32]
#> 67: (58,62,36), [9,79,68]
#> 68: (58,4,62), [56,67,82]
#> 69: (75,51,29), [64,117,75]
#> 70: (100,47,67), [108,41,99]
#> 71: (81,11,15), [55,53,111]
#> 72: (19,99,17), [85,94,93]
#> 73: (61,26,92), [125,136,130]
#> 74: (51,82,26), [125,24,80]
#> 75: (46,51,75), [69,109,80]
#> 76: (67,34,29), [98,64,108]
#> 77: (7,96,100), [99,50,66]
#> 78: (58,14,44), [106,81,45]
#> 79: (33,58,36), [67,92,45]
#> 80: (46,82,51), [74,75,124]
#> 81: (58,44,95), [84,82,78]
#> 82: (58,95,4), [83,68,81]
#> 83: (4,95,16), [152,56,82]
#> 84: (84,95,44), [81,106,0]
#> 85: (99,64,17), [128,72,65]
#> 86: (78,64,52), [89,129,128]
#> 87: (24,48,49), [47,169,95]
#> 88: (49,54,23), [48,59,47]
#> 89: (52,64,26), [33,130,86]
#> 90: (74,81,99), [53,93,102]
#> 91: (14,33,84), [113,106,45]
#> 92: (33,36,48), [46,114,79]
#> 93: (19,74,99), [90,72,103]
#> 94: (17,78,19), [127,72,128]
#> 95: (8,48,24), [87,168,114]
#> 96: (72,49,40), [137,170,169]
#> 97: (28,29,3), [98,133,117]
#> 98: (29,34,3), [110,97,76]
#> 99: (100,96,47), [150,70,77]
#> 100: (37,81,27), [101,112,111]
#> 101: (81,86,27), [123,100,102]
#> 102: (86,81,74), [90,107,101]
#> 103: (94,74,19), [93,126,107]
#> 104: (12,11,37), [111,140,58]
#> 105: (32,86,94), [107,147,123]
#> 106: (84,44,14), [78,91,84]
#> 107: (94,86,74), [102,103,105]
#> 108: (67,47,34), [110,76,70]
#> 109: (75,68,46), [115,75,116]
#> 110: (3,34,47), [108,145,98]
#> 111: (37,11,81), [71,100,104]
#> 112: (27,32,37), [162,100,123]
#> 113: (8,84,33), [91,114,0]
#> 114: (8,33,48), [92,95,113]
#> 115: (10,46,68), [109,142,124]
#> 116: (75,28,68), [132,109,117]
#> 117: (28,75,29), [69,97,116]
#> 118: (92,82,10), [124,141,125]
#> 119: (20,40,53), [137,60,120]
#> 120: (20,12,40), [143,119,121]
#> 121: (12,20,89), [51,58,120]
#> 122: (45,37,63), [162,154,140]
#> 123: (32,27,86), [101,105,112]
#> 124: (10,82,46), [80,115,118]
#> 125: (26,82,92), [118,73,74]
#> 126: (42,94,19), [103,159,156]
#> 127: (19,78,93), [160,173,94]
#> 128: (78,17,64), [85,86,94]
#> 129: (52,83,78), [148,86,135]
#> 130: (61,52,26), [89,73,138]
#> 131: (68,59,85), [144,142,132]
#> 132: (59,68,28), [116,133,131]
#> 133: (59,28,3), [97,134,132]
#> 134: (3,79,59), [144,133,145]
#> 135: (52,88,83), [149,129,138]
#> 136: (92,85,61), [188,73,141]
#> 137: (40,49,53), [59,119,96]
#> 138: (52,61,88), [179,135,130]
#> 139: (12,45,43), [153,143,140]
#> 140: (37,45,12), [139,104,122]
#> 141: (92,10,85), [142,136,118]
#> 142: (68,85,10), [141,115,131]
#> 143: (12,43,40), [171,120,139]
#> 144: (79,85,59), [131,134,0]
#> 145: (3,47,79), [146,134,110]
#> 146: (47,13,79), [0,145,150]
#> 147: (94,39,32), [157,105,156]
#> 148: (35,78,83), [129,149,160]
#> 149: (35,83,88), [135,161,148]
#> 150: (47,96,13), [151,146,99]
#> 151: (13,96,16), [66,152,150]
#> 152: (13,16,95), [83,0,151]
#> 153: (98,43,45), [139,154,175]
#> 154: (45,63,98), [176,153,122]
#> 155: (63,32,70), [157,163,162]
#> 156: (42,39,94), [147,126,172]
#> 157: (32,39,70), [163,155,147]
#> 158: (93,35,31), [166,173,160]
#> 159: (19,31,42), [164,126,173]
#> 160: (35,93,78), [127,148,158]
#> 161: (60,35,88), [149,179,178]
#> 162: (37,32,63), [155,122,112]
#> 163: (63,70,39), [157,181,155]
#> 164: (31,21,42), [183,159,165]
#> 165: (31,65,21), [177,164,166]
#> 166: (31,35,65), [167,165,158]
#> 167: (35,76,65), [177,166,178]
#> 168: (72,8,24), [95,169,0]
#> 169: (72,24,49), [87,96,168]
#> 170: (66,72,40), [96,171,0]
#> 171: (66,40,43), [143,175,170]
#> 172: (39,42,56), [182,180,156]
#> 173: (19,93,31), [158,159,127]
#> 174: (55,21,76), [177,186,183]
#> 175: (43,98,66), [176,171,153]
#> 176: (66,98,63), [154,184,175]
#> 177: (21,65,76), [167,174,165]
#> 178: (60,76,35), [167,161,187]
#> 179: (88,61,60), [188,161,138]
#> 180: (39,56,90), [185,181,172]
#> 181: (39,90,63), [184,163,180]
#> 182: (42,55,56), [185,172,183]
#> 183: (42,21,55), [174,182,164]
#> 184: (63,90,66), [0,176,181]
#> 185: (90,56,55), [182,186,180]
#> 186: (90,55,76), [174,187,185]
#> 187: (60,90,76), [186,178,0]
#> 188: (61,85,60), [0,179,136]
#> boundary nodes:  84 95 13 79 85 60 90 66 72 8 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.106175349 0.134040331 0.045803768 0.062500906 0.156549527 0.117789047 
#>                                                                         
#> 0.040870850 0.032007081 0.157969706 0.209712009 0.095484975 0.119849214 
#>                                                                         
#> 0.203497883 0.115871555 0.045464167 0.144198892 0.045979938 0.098979205 
#>                                                                         
#> 0.024084685 0.167337382 0.080800368 0.157004096 0.200937886 0.072663474 
#>                                                                         
#> 0.063375182 0.138196503 0.107824325 0.252701288 0.172077299 0.072707930 
#>                                                                         
#> 0.105543052 0.074643037 0.154514876 0.133990302 0.050936943 0.146908160 
#>                                                                         
#> 0.051181794 0.090401094 0.181407894 0.049858868 0.054501088 0.016514142 
#>                                                                         
#> 0.135730320 0.066615005 0.053307520 0.092291078 0.198875997 0.058684841 
#>                                                                         
#> 0.157291510 0.101663924 0.059104687 0.048113133 0.218703771 0.261745038 
#>                                                                         
#> 0.244484557 0.180976058 0.169416257 0.122263629 0.144068582 0.089983903 
#>                                                                         
#> 0.049033302 0.089272981 0.092083427 0.093991954 0.152841599 0.098518406 
#>                                                                         
#> 0.022923394 0.078475624 0.130073093 0.130393236 0.069341934 0.140226618 
#>                                                                         
#> 0.083886344 0.092095723 0.081170172 0.104027691 0.130346328 0.157604189 
#>                                                                         
#> 0.322506618 0.099558661 0.024746298 0.196827977 0.054659591 0.008631949 
#>                                                                         
#> 0.046770101 0.079568356 0.213676735 0.062355334 0.049894622 0.066698655 
#>                                                                         
#> 0.073573393 0.234645611 0.056363394 0.204227956 0.048280053 0.064445566 
#>                                                                         
#> 0.090159996 0.179885225 0.175587655 0.124045897 0.059749387 0.100104807 
#>                                                                         
#> 0.072642881 0.082618522 0.036361154 0.171732545 0.171772540 0.081287570 
#>                                                                         
#> 0.103084462 0.107684764 0.152654841 0.154497107 0.072084311 0.078365390 
#>                                                                         
#> 0.046980355 0.182719440 0.098582552 0.093611805 0.105548733 0.237196421 
#>                                                                         
#> 0.136888102 0.087885237 0.140015067 0.045791882 0.054381058 0.135797232 
#>                                                                         
#> 0.178651242 0.129662807 0.119635142 0.036123339 0.037701868 0.078044010 
#>                                                                         
#> 0.131638956 0.038652307 0.117010830 0.067313663 0.022186879 0.104735001 
#>                                                                         
#> 0.147420153 0.089966193 0.374001548 0.019406319 0.353864822 0.019252041 
#>                                                                         
#> 0.106689629 0.622367828 0.153689905 0.163780988 0.523903320 0.124011601 
#>                                                                         
#> 0.044592856 0.131077182 0.142810459 0.207566786 0.036011211 0.165991359 
#>                                                                         
#> 0.191847338 0.159399206 0.170611114 0.018762768 0.218299231 0.193362274 
#>                                                                         
#> 0.136570838 0.134881100 0.071213875 0.247802421 0.070106424 0.072598956 
#>                                                                         
#> 0.259505951 0.161793202 0.051557003 0.102540628 0.034816863 0.103868769 
#>                                                                         
#> 0.028266743 0.104525283 0.192379920 0.154122121 0.144861733 0.060788256 
#>                                                                         
#> 0.083541135 0.115139034 0.119645994 0.183077288 0.154126104 0.142239235 
#>                                                                         
#> 0.080394690 0.064142364 0.066769844 0.135344584 0.159017526 0.176426398 
#>                                                                         
#> 0.187332160 0.052696207 0.086401976 0.018347367 0.128958668 0.084970819 
#>                                                                         
#> 0.250631274 0.062554910 0.098517729 0.118004442 0.159594326 0.124811321 
#>                                                                         
#> 0.075404934 0.125610491 0.217852344 0.088201558 0.188371050 0.109514075 
#>                                                                         
#> 0.043351457 0.163855927 0.077666510 0.117473259 0.151248491 0.172197894 
#>                                                                         
#> 0.097226951 0.076379178 0.089760112 0.045450814 0.166698029 0.187777767 
#>                                                                         
#> 0.147496612 0.059206347 0.103321582 0.186732873 0.147665260 0.144706830 
#>                                                                         
#> 0.160170048 0.101323320 0.044253894 0.067577210 0.103385023 0.094972892 
#>                                                                         
#> 0.058562028 0.076603680 0.087987888 0.105297573 0.101277856 0.498648278 
#>                                                                         
#> 0.628675060 0.035474292 0.086713191 0.166963197 0.101620714 0.235728908 
#>                                                                         
#> 0.131431978 0.028966345 0.039196090 0.392315849 0.374739269 0.225608049 
#>                                                                         
#> 0.119436730 0.207979641 0.160267429 0.268838947 0.111779555 0.226747344 
#>                                                                         
#> 0.078533606 0.016648174 0.036063254 0.151648691 0.170931589 0.060819874 
#>                                                                         
#> 0.131433060 0.092358347 0.064168874 0.172400645 0.124909984 0.184929862 
#>                                                                         
#> 0.181522891 0.103218846 0.051318619 0.051465247 0.230204162 0.084482076 
#>                                                                         
#> 0.203710228 0.148095742 0.185958317 0.082451781 0.270966588 0.080503135 
#>                                                             
#> 0.149808884 0.052619146 0.522554640 0.228592016 0.196521019 
#> 
#> $alpha
#> [1] 0.2965856
#> 
#> attr(,"class")
#> [1] "scree"