Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2           mx1
#>    75   88 0.993315194 0.29591132 0.935886916 0.19136535   1.062958390
#>    88   71 0.935886916 0.19136535 0.814911766 0.16073303   0.882282862
#>    71   65 0.814911766 0.16073303 0.817811489 0.03413717   0.872086439
#>    65    7 0.817811489 0.03413717 0.948466608 0.00000000   0.896166272
#>     7   75 0.948466608 0.00000000 0.993315194 0.29591132   1.520110683
#>    74   13 0.832325634 0.99872244 0.741879841 1.00000000   0.787429234
#>    13   40 0.741879841 1.00000000 0.782600637 0.95224338   0.778301036
#>    40   22 0.782600637 0.95224338 0.792323818 0.92200363   0.815540571
#>    22   58 0.792323818 0.92200363 0.835014442 0.89388376   0.834088159
#>    58   74 0.835014442 0.89388376 0.832325634 0.99872244   0.869027792
#>    51    7 0.339707884 0.01966239 0.948466608 0.00000000   0.592572751
#>    65   82 0.817811489 0.03413717 0.714862314 0.10860981   0.702794812
#>    82   77 0.714862314 0.10860981 0.629346647 0.12673091   0.660014388
#>    77   16 0.629346647 0.12673091 0.581947739 0.08344711   0.636744672
#>    16   24 0.581947739 0.08344711 0.470143752 0.09675557   0.521173193
#>    24   51 0.470143752 0.09675557 0.339707884 0.01966239   0.471645337
#>    47   66 0.745852145 0.73465893 0.680130470 0.66359757   0.738616268
#>    66   42 0.680130470 0.66359757 0.726379911 0.54945389   0.734990402
#>    42   21 0.726379911 0.54945389 0.909032880 0.58736357   0.818254213
#>    21   95 0.909032880 0.58736357 0.827747649 0.70072114   0.863317912
#>    95   20 0.827747649 0.70072114 0.809436900 0.69245114   0.834391679
#>    20   47 0.809436900 0.69245114 0.745852145 0.73465893   0.762124031
#>    54   69 0.001047047 0.51156716 0.052906485 0.51021096   0.028804487
#>    69   10 0.052906485 0.51021096 0.255300466 0.63532707   0.141260929
#>    10    8 0.255300466 0.63532707 0.215220227 0.70879959   0.143829961
#>     8   34 0.215220227 0.70879959 0.179356566 0.75290692   0.128053233
#>    34   12 0.179356566 0.75290692 0.072975889 0.86220625   0.072609990
#>    12   54 0.072975889 0.86220625 0.001047047 0.51156716  -2.540278036
#>    50   28 0.343683136 0.41913482 0.157827481 0.43865150   0.249044303
#>    28   68 0.157827481 0.43865150 0.168377924 0.33592683   0.230648833
#>    68   17 0.168377924 0.33592683 0.239531091 0.26328642   0.219838868
#>    17   52 0.239531091 0.26328642 0.272786371 0.29852971   0.219879183
#>    52  100 0.272786371 0.29852971 0.288230628 0.28355627   0.296699456
#>   100   43 0.288230628 0.28355627 0.417148803 0.34502818   0.347347991
#>    43   50 0.417148803 0.34502818 0.343683136 0.41913482   0.343945149
#>    83   75 1.000000000 0.31791526 0.993315194 0.29591132   2.066438855
#>    51   46 0.339707884 0.01966239 0.024273464 0.02970712 -32.189053259
#>    10   33 0.255300466 0.63532707 0.296192583 0.70453352   0.254434392
#>    33   34 0.296192583 0.70453352 0.179356566 0.75290692   0.259652269
#>    58   23 0.835014442 0.89388376 0.815471781 0.83884356   0.873826400
#>    23   95 0.815471781 0.83884356 0.827747649 0.70072114   0.898722034
#>    95   76 0.827747649 0.70072114 0.925047765 0.59903264   0.968263257
#>    76   79 0.925047765 0.59903264 0.994328469 0.46191417   1.366780127
#>    69   62 0.052906485 0.51021096 0.269423485 0.56564748   0.150930505
#>    62   14 0.269423485 0.56564748 0.260710138 0.63366912   0.246602795
#>    14   33 0.260710138 0.63366912 0.296192583 0.70453352   0.269919672
#>    80   56 0.716714030 0.93012827 0.524078245 0.93268789   0.621176365
#>    56   30 0.524078245 0.93268789 0.547317362 0.81571581   0.474334385
#>    30   80 0.547317362 0.81571581 0.716714030 0.93012827   0.644443099
#>    34   59 0.179356566 0.75290692 0.153506756 0.86794090   0.117909411
#>    59   97 0.153506756 0.86794090 0.105593741 0.89629217   0.113909909
#>    97   72 0.105593741 0.89629217 0.290793858 0.92393505  -2.321114181
#>    72   56 0.290793858 0.92393505 0.524078245 0.93268789   0.374646611
#>    80   40 0.716714030 0.93012827 0.782600637 0.95224338   0.743340214
#>    62   25 0.269423485 0.56564748 0.285422201 0.61744025   0.255232212
#>    25   33 0.285422201 0.61744025 0.296192583 0.70453352   0.295919947
#>    85   71 0.758823029 0.12540050 0.814911766 0.16073303   0.767768851
#>    71   57 0.814911766 0.16073303 0.821921516 0.19969919   0.801041496
#>    57   64 0.821921516 0.19969919 0.975370795 0.33763643   0.860696839
#>    64   49 0.975370795 0.33763643 0.951443658 0.40827191   0.844740504
#>    49   32 0.951443658 0.40827191 0.739999629 0.50744435   0.802239457
#>    32   36 0.739999629 0.50744435 0.743067006 0.21724436   0.790952559
#>    36   85 0.743067006 0.21724436 0.758823029 0.12540050   0.766523781
#>    77   38 0.629346647 0.12673091 0.574206970 0.16889953   0.585758040
#>    38   24 0.574206970 0.16889953 0.470143752 0.09675557   0.529819175
#>    68   31 0.168377924 0.33592683 0.163791134 0.25718830   0.198890699
#>    31   48 0.163791134 0.25718830 0.165254442 0.25535210   0.199722032
#>    48   45 0.165254442 0.25535210 0.210389579 0.23463513   0.200883331
#>    45   81 0.210389579 0.23463513 0.251726532 0.17101736   0.255014471
#>    81  100 0.251726532 0.17101736 0.288230628 0.28355627   0.285128470
#>    70   46 0.009417436 0.42162507 0.024273464 0.02970712  -0.254818794
#>    46   55 0.024273464 0.02970712 0.078154732 0.16308367  -0.238569540
#>    55   31 0.078154732 0.16308367 0.163791134 0.25718830   0.078991858
#>    28   70 0.157827481 0.43865150 0.009417436 0.42162507   0.086459147
#>    70   27 0.009417436 0.42162507 0.005936734 0.49034881  -0.012738648
#>    27   54 0.005936734 0.49034881 0.001047047 0.51156716   0.007349032
#>    95   47 0.827747649 0.70072114 0.745852145 0.73465893   0.797451671
#>    42   32 0.726379911 0.54945389 0.739999629 0.50744435   0.820148656
#>    49   79 0.951443658 0.40827191 0.994328469 0.46191417   0.902610794
#>    80   93 0.716714030 0.93012827 0.763749913 0.87347464   0.753055622
#>    93   73 0.763749913 0.87347464 0.807165326 0.86241903   0.791087393
#>    73   58 0.807165326 0.86241903 0.835014442 0.89388376   0.804005463
#>    73   92 0.807165326 0.86241903 0.748597204 0.83550917   0.781880774
#>    92   47 0.748597204 0.83550917 0.745852145 0.73465893   0.784682674
#>    38   61 0.574206970 0.16889953 0.422176041 0.11360812   0.478172284
#>    61   51 0.422176041 0.11360812 0.339707884 0.01966239   0.420436763
#>    62   53 0.269423485 0.56564748 0.341348811 0.57345371   0.303886368
#>    53   33 0.341348811 0.57345371 0.296192583 0.70453352   0.358362347
#>    72   33 0.290793858 0.92393505 0.296192583 0.70453352   0.284361911
#>    33   30 0.296192583 0.70453352 0.547317362 0.81571581   0.406407546
#>    70   44 0.009417436 0.42162507 0.030617664 0.47813810   0.004210634
#>    44   69 0.030617664 0.47813810 0.052906485 0.51021096   0.027942087
#>    50   62 0.343683136 0.41913482 0.269423485 0.56564748   0.254661322
#>    51   81 0.339707884 0.01966239 0.251726532 0.17101736   0.219348887
#>    81   55 0.251726532 0.17101736 0.078154732 0.16308367   0.165482198
#>    33   35 0.296192583 0.70453352 0.551549771 0.72987317   0.418992938
#>    35   96 0.551549771 0.72987317 0.629363692 0.73497964   0.587683846
#>    96   92 0.629363692 0.73497964 0.748597204 0.83550917   0.647034693
#>    92   80 0.748597204 0.83550917 0.716714030 0.93012827   0.645228754
#>    55   48 0.078154732 0.16308367 0.165254442 0.25535210   0.114277591
#>    42   90 0.726379911 0.54945389 0.670080345 0.53176755   0.704928967
#>    90   78 0.670080345 0.53176755 0.656577786 0.50544486   0.698428760
#>    78   36 0.656577786 0.50544486 0.743067006 0.21724436   0.701752638
#>    53   89 0.341348811 0.57345371 0.538981067 0.67979398   0.415566762
#>    89   35 0.538981067 0.67979398 0.551549771 0.72987317   0.422029470
#>    72   34 0.290793858 0.92393505 0.179356566 0.75290692   0.272970721
#>    72   59 0.290793858 0.92393505 0.153506756 0.86794090   0.249421952
#>    89   96 0.538981067 0.67979398 0.629363692 0.73497964   0.593054583
#>    78   91 0.656577786 0.50544486 0.547788648 0.37643860   0.697425476
#>    91   39 0.547788648 0.37643860 0.572208237 0.27592023   0.696348124
#>    39   84 0.572208237 0.27592023 0.617233743 0.21224922   0.676684893
#>    84   87 0.617233743 0.21224922 0.695517016 0.17457066   0.679121377
#>    87   85 0.695517016 0.17457066 0.758823029 0.12540050   0.742635243
#>    94   43 0.497906548 0.32680723 0.417148803 0.34502818   0.435466773
#>    61   94 0.422176041 0.11360812 0.497906548 0.32680723   0.450682202
#>    77   84 0.629346647 0.12673091 0.617233743 0.21224922   0.617753074
#>    39   94 0.572208237 0.27592023 0.497906548 0.32680723   0.479789378
#>    91   94 0.547788648 0.37643860 0.497906548 0.32680723   0.550467081
#>    63   50 0.360223063 0.48246668 0.343683136 0.41913482   0.389397303
#>    50   86 0.343683136 0.41913482 0.421888423 0.37202791   0.403356504
#>    86   67 0.421888423 0.37202791 0.552709237 0.49615787   0.478919373
#>    67   60 0.552709237 0.49615787 0.423222070 0.52312648   0.478160187
#>    60   53 0.423222070 0.52312648 0.341348811 0.57345371   0.372914079
#>    53   63 0.341348811 0.57345371 0.360223063 0.48246668   0.371950480
#>    66   98 0.680130470 0.66359757 0.675019002 0.54180630   0.692293496
#>    98   90 0.675019002 0.54180630 0.670080345 0.53176755   0.704345298
#>    94   86 0.497906548 0.32680723 0.421888423 0.37202791   0.460986704
#>    96   66 0.629363692 0.73497964 0.680130470 0.66359757   0.687574374
#>    64   79 0.975370795 0.33763643 0.994328469 0.46191417   1.024659017
#>    63   62 0.360223063 0.48246668 0.269423485 0.56564748   0.267838152
#>    60   99 0.423222070 0.52312648 0.502505158 0.56643043   0.423936329
#>    99   89 0.502505158 0.56643043 0.538981067 0.67979398   0.425578612
#>    88   64 0.935886916 0.19136535 0.975370795 0.33763643   0.885335174
#>    67   99 0.552709237 0.49615787 0.502505158 0.56643043   0.486333958
#>    83   79 1.000000000 0.31791526 0.994328469 0.46191417   1.038753013
#>    85   82 0.758823029 0.12540050 0.714862314 0.10860981   0.758440713
#>    89   98 0.538981067 0.67979398 0.675019002 0.54180630   0.602002897
#>    86   91 0.421888423 0.37202791 0.547788648 0.37643860   0.482566745
#>    98   99 0.675019002 0.54180630 0.502505158 0.56643043   0.592078640
#>    93   92 0.763749913 0.87347464 0.748597204 0.83550917   0.779645778
#>    82   87 0.714862314 0.10860981 0.695517016 0.17457066   0.675316641
#>     1   88 0.948940048 0.13348414 0.935886916 0.19136535   1.062958390
#>     1   71 0.948940048 0.13348414 0.814911766 0.16073303   0.882282862
#>     1   65 0.948940048 0.13348414 0.817811489 0.03413717   0.872086439
#>     1    7 0.948940048 0.13348414 0.948466608 0.00000000   0.896166272
#>     1   75 0.948940048 0.13348414 0.993315194 0.29591132   1.520110683
#>     2   13 0.814724366 0.97964024 0.741879841 1.00000000   0.787429234
#>     2   40 0.814724366 0.97964024 0.782600637 0.95224338   0.778301036
#>     2   22 0.814724366 0.97964024 0.792323818 0.92200363   0.815540571
#>     2   58 0.814724366 0.97964024 0.835014442 0.89388376   0.834088159
#>     2   74 0.814724366 0.97964024 0.832325634 0.99872244   0.869027792
#>     3    7 0.590270586 0.03946092 0.948466608 0.00000000   0.592572751
#>     3   65 0.590270586 0.03946092 0.817811489 0.03413717   0.685918679
#>     3   82 0.590270586 0.03946092 0.714862314 0.10860981   0.702794812
#>     3   77 0.590270586 0.03946092 0.629346647 0.12673091   0.660014388
#>     3   16 0.590270586 0.03946092 0.581947739 0.08344711   0.636744672
#>     3   24 0.590270586 0.03946092 0.470143752 0.09675557   0.521173193
#>     3   51 0.590270586 0.03946092 0.339707884 0.01966239   0.471645337
#>     4   66 0.795090016 0.65616011 0.680130470 0.66359757   0.738616268
#>     4   42 0.795090016 0.65616011 0.726379911 0.54945389   0.734990402
#>     4   21 0.795090016 0.65616011 0.909032880 0.58736357   0.818254213
#>     4   95 0.795090016 0.65616011 0.827747649 0.70072114   0.863317912
#>     4   20 0.795090016 0.65616011 0.809436900 0.69245114   0.834391679
#>     4   47 0.795090016 0.65616011 0.745852145 0.73465893   0.762124031
#>     5   69 0.036476911 0.65495361 0.052906485 0.51021096   0.028804487
#>     5   10 0.036476911 0.65495361 0.255300466 0.63532707   0.141260929
#>     5    8 0.036476911 0.65495361 0.215220227 0.70879959   0.143829961
#>     5   34 0.036476911 0.65495361 0.179356566 0.75290692   0.128053233
#>     5   12 0.036476911 0.65495361 0.072975889 0.86220625   0.072609990
#>     5   54 0.036476911 0.65495361 0.001047047 0.51156716  -2.540278036
#>     6   28 0.274984116 0.32135059 0.157827481 0.43865150   0.249044303
#>     6   68 0.274984116 0.32135059 0.168377924 0.33592683   0.230648833
#>     6   17 0.274984116 0.32135059 0.239531091 0.26328642   0.219838868
#>     6   52 0.274984116 0.32135059 0.272786371 0.29852971   0.219879183
#>     6  100 0.274984116 0.32135059 0.288230628 0.28355627   0.296699456
#>     6   43 0.274984116 0.32135059 0.417148803 0.34502818   0.347347991
#>     6   50 0.274984116 0.32135059 0.343683136 0.41913482   0.343945149
#>     7   46 0.948466608 0.00000000 0.024273464 0.02970712 -32.189053259
#>     8   33 0.215220227 0.70879959 0.296192583 0.70453352   0.254434392
#>     9   58 0.926174804 0.87686932 0.835014442 0.89388376   0.892246533
#>     9   23 0.926174804 0.87686932 0.815471781 0.83884356   0.873826400
#>     9   95 0.926174804 0.87686932 0.827747649 0.70072114   0.898722034
#>     9   76 0.926174804 0.87686932 0.925047765 0.59903264   0.968263257
#>     9   79 0.926174804 0.87686932 0.994328469 0.46191417   1.366780127
#>    10   62 0.255300466 0.63532707 0.269423485 0.56564748   0.150930505
#>    10   14 0.255300466 0.63532707 0.260710138 0.63366912   0.246602795
#>    11   56 0.551103051 0.90171325 0.524078245 0.93268789   0.621176365
#>    11   30 0.551103051 0.90171325 0.547317362 0.81571581   0.474334385
#>    11   80 0.551103051 0.90171325 0.716714030 0.93012827   0.644443099
#>    12   59 0.072975889 0.86220625 0.153506756 0.86794090   0.117909411
#>    12   97 0.072975889 0.86220625 0.105593741 0.89629217   0.113909909
#>    13   72 0.741879841 1.00000000 0.290793858 0.92393505  -2.321114181
#>    13   56 0.741879841 1.00000000 0.524078245 0.93268789   0.374646611
#>    13   80 0.741879841 1.00000000 0.716714030 0.93012827   0.621359895
#>    14   25 0.260710138 0.63366912 0.285422201 0.61744025   0.255232212
#>    15   71 0.800212368 0.20985315 0.814911766 0.16073303   0.767768851
#>    15   57 0.800212368 0.20985315 0.821921516 0.19969919   0.801041496
#>    15   64 0.800212368 0.20985315 0.975370795 0.33763643   0.860696839
#>    15   49 0.800212368 0.20985315 0.951443658 0.40827191   0.844740504
#>    15   32 0.800212368 0.20985315 0.739999629 0.50744435   0.802239457
#>    15   36 0.800212368 0.20985315 0.743067006 0.21724436   0.790952559
#>    15   85 0.800212368 0.20985315 0.758823029 0.12540050   0.766523781
#>    16   38 0.581947739 0.08344711 0.574206970 0.16889953   0.585758040
#>    17   31 0.239531091 0.26328642 0.163791134 0.25718830   0.198890699
#>    17   48 0.239531091 0.26328642 0.165254442 0.25535210   0.199722032
#>    17   45 0.239531091 0.26328642 0.210389579 0.23463513   0.200883331
#>    17   81 0.239531091 0.26328642 0.251726532 0.17101736   0.255014471
#>    17  100 0.239531091 0.26328642 0.288230628 0.28355627   0.285128470
#>    18   46 0.059670771 0.33138096 0.024273464 0.02970712  -0.254818794
#>    18   55 0.059670771 0.33138096 0.078154732 0.16308367  -0.238569540
#>    18   31 0.059670771 0.33138096 0.163791134 0.25718830   0.078991858
#>    18   68 0.059670771 0.33138096 0.168377924 0.33592683   0.115452287
#>    18   28 0.059670771 0.33138096 0.157827481 0.43865150   0.112000926
#>    18   70 0.059670771 0.33138096 0.009417436 0.42162507   0.086459147
#>    19   70 0.000000000 0.49289189 0.009417436 0.42162507  -2.457716122
#>    19   27 0.000000000 0.49289189 0.005936734 0.49034881  -0.012738648
#>    19   54 0.000000000 0.49289189 0.001047047 0.51156716   0.007349032
#>    21   32 0.909032880 0.58736357 0.739999629 0.50744435   0.820148656
#>    21   49 0.909032880 0.58736357 0.951443658 0.40827191   0.856245288
#>    21   79 0.909032880 0.58736357 0.994328469 0.46191417   0.902610794
#>    21   76 0.909032880 0.58736357 0.925047765 0.59903264   0.961921677
#>    22   80 0.792323818 0.92200363 0.716714030 0.93012827   0.754570688
#>    22   93 0.792323818 0.92200363 0.763749913 0.87347464   0.753055622
#>    22   73 0.792323818 0.92200363 0.807165326 0.86241903   0.791087393
#>    23   92 0.815471781 0.83884356 0.748597204 0.83550917   0.781880774
#>    23   47 0.815471781 0.83884356 0.745852145 0.73465893   0.784682674
#>    23   73 0.815471781 0.83884356 0.807165326 0.86241903   0.840547492
#>    24   61 0.470143752 0.09675557 0.422176041 0.11360812   0.478172284
#>    25   53 0.285422201 0.61744025 0.341348811 0.57345371   0.303886368
#>    26   72 0.329266651 0.91455083 0.290793858 0.92393505   0.394745259
#>    26   33 0.329266651 0.91455083 0.296192583 0.70453352   0.284361911
#>    26   30 0.329266651 0.91455083 0.547317362 0.81571581   0.406407546
#>    26   56 0.329266651 0.91455083 0.524078245 0.93268789   0.433169655
#>    27   44 0.005936734 0.49034881 0.030617664 0.47813810   0.004210634
#>    27   69 0.005936734 0.49034881 0.052906485 0.51021096   0.027942087
#>    28   44 0.157827481 0.43865150 0.030617664 0.47813810   0.094049626
#>    28   62 0.157827481 0.43865150 0.269423485 0.56564748   0.254661322
#>    28   69 0.157827481 0.43865150 0.052906485 0.51021096   0.157755855
#>    29   46 0.102683814 0.09400938 0.024273464 0.02970712   0.027199929
#>    29   51 0.102683814 0.09400938 0.339707884 0.01966239   0.178700284
#>    29   81 0.102683814 0.09400938 0.251726532 0.17101736   0.219348887
#>    29   55 0.102683814 0.09400938 0.078154732 0.16308367   0.165482198
#>    30   35 0.547317362 0.81571581 0.551549771 0.72987317   0.418992938
#>    30   96 0.547317362 0.81571581 0.629363692 0.73497964   0.587683846
#>    30   92 0.547317362 0.81571581 0.748597204 0.83550917   0.647034693
#>    32   90 0.739999629 0.50744435 0.670080345 0.53176755   0.704928967
#>    32   78 0.739999629 0.50744435 0.656577786 0.50544486   0.698428760
#>    33   89 0.296192583 0.70453352 0.538981067 0.67979398   0.415566762
#>    36   91 0.743067006 0.21724436 0.547788648 0.37643860   0.697425476
#>    36   39 0.743067006 0.21724436 0.572208237 0.27592023   0.696348124
#>    36   84 0.743067006 0.21724436 0.617233743 0.21224922   0.676684893
#>    36   87 0.743067006 0.21724436 0.695517016 0.17457066   0.679121377
#>    37   43 0.349413480 0.17214600 0.417148803 0.34502818   0.435466773
#>    37  100 0.349413480 0.17214600 0.288230628 0.28355627   0.378325759
#>    37   81 0.349413480 0.17214600 0.251726532 0.17101736   0.300039067
#>    37   51 0.349413480 0.17214600 0.339707884 0.01966239   0.301412623
#>    37   61 0.349413480 0.17214600 0.422176041 0.11360812   0.347837059
#>    37   94 0.349413480 0.17214600 0.497906548 0.32680723   0.450682202
#>    38   84 0.574206970 0.16889953 0.617233743 0.21224922   0.617753074
#>    38   39 0.574206970 0.16889953 0.572208237 0.27592023   0.563822514
#>    38   94 0.574206970 0.16889953 0.497906548 0.32680723   0.479789378
#>    41   50 0.401239974 0.49030226 0.343683136 0.41913482   0.389397303
#>    41   86 0.401239974 0.49030226 0.421888423 0.37202791   0.403356504
#>    41   67 0.401239974 0.49030226 0.552709237 0.49615787   0.478919373
#>    41   60 0.401239974 0.49030226 0.423222070 0.52312648   0.478160187
#>    41   53 0.401239974 0.49030226 0.341348811 0.57345371   0.372914079
#>    41   63 0.401239974 0.49030226 0.360223063 0.48246668   0.371950480
#>    42   98 0.726379911 0.54945389 0.675019002 0.54180630   0.692293496
#>    43   86 0.417148803 0.34502818 0.421888423 0.37202791   0.460986704
#>    45   55 0.210389579 0.23463513 0.078154732 0.16308367   0.156377641
#>    47   96 0.745852145 0.73465893 0.629363692 0.73497964   0.687750762
#>    53   99 0.341348811 0.57345371 0.502505158 0.56643043   0.423936329
#>    57   88 0.821921516 0.19969919 0.935886916 0.19136535   0.877013372
#>    64   75 0.975370795 0.33763643 0.993315194 0.29591132   0.898594292
#>    64   83 0.975370795 0.33763643 1.000000000 0.31791526   0.976004369
#>    65   85 0.817811489 0.03413717 0.758823029 0.12540050   0.815623259
#>    66   89 0.680130470 0.66359757 0.538981067 0.67979398   0.608987636
#>    67   91 0.552709237 0.49615787 0.547788648 0.37643860   0.482566745
#>    67   78 0.552709237 0.49615787 0.656577786 0.50544486   0.610632703
#>    67   90 0.552709237 0.49615787 0.670080345 0.53176755   0.600154074
#>    67   98 0.552709237 0.49615787 0.675019002 0.54180630   0.592525914
#>    77   87 0.629346647 0.12673091 0.695517016 0.17457066   0.675316641
#>     7   83 0.948466608 0.00000000 1.000000000 0.31791526   2.066438855
#>     9   74 0.926174804 0.87686932 0.832325634 0.99872244   0.892246533
#>    13   97 0.741879841 1.00000000 0.105593741 0.89629217  -2.321114181
#>    19   46 0.000000000 0.49289189 0.024273464 0.02970712  -2.457716122
#>            my1           mx2           my2 bp1 bp2
#>   1.896095e-01   0.898594292  2.798966e-01   0   0
#>   1.488643e-01   0.877013372  1.696750e-01   0   0
#>   9.871150e-02   0.815623259  9.741819e-02   0   0
#>   6.692841e-02   0.685918679 -7.377640e-01   0   0
#>   6.471541e-02   2.066438855 -1.808658e-02   0   0
#>   1.022476e+00   0.943570513  1.207662e+01   0   1
#>   9.898163e-01   0.743340214  9.600061e-01   0   0
#>   9.461517e-01   0.754570688  9.265477e-01   0   0
#>   9.389432e-01   0.804005463  8.932726e-01   0   0
#>   9.472099e-01   0.892246533  9.478054e-01   0   0
#>  -1.585087e+00 -32.189053259 -1.016523e+03   0   0
#>  -1.646555e-02   0.758440713  6.045799e-02   0   0
#>   6.061575e-02   0.675316641  1.328289e-01   0   0
#>   7.103501e-02   0.585758040  1.268691e-01   0   0
#>   4.916716e-02   0.529819175  1.218018e-01   0   0
#>  -5.467543e-02   0.420436763  3.196563e-02   0   0
#>   6.754288e-01   0.687574374  7.226353e-01   0   0
#>   6.193844e-01   0.692293496  6.020842e-01   0   0
#>   5.657693e-01   0.820148656  5.566417e-01   0   0
#>   6.404051e-01   0.876011339  6.495072e-01   0   0
#>   6.616044e-01   0.797451671  7.433939e-01   0   0
#>   6.901739e-01   0.797451671  7.433939e-01   0   0
#>   5.807789e-01   0.026857869  5.063426e-01   0   0
#>   5.935437e-01   0.150930505  5.779018e-01   0   0
#>   6.221868e-01   0.254434392  6.825230e-01   0   0
#>   6.745582e-01   0.259652269  7.815613e-01   0   0
#>   7.554305e-01   0.117909411  7.995203e-01   0   0
#>   1.215583e+00  -5.277047448  1.776994e+00   0   1
#>   4.125994e-01   0.254661322  4.660898e-01   0   0
#>   3.942266e-01   0.112000926  3.820407e-01   0   0
#>   3.151658e-01   0.198890699  2.946465e-01   0   0
#>   3.151412e-01   0.263743652  2.737510e-01   0   0
#>   3.077430e-01   0.263743652  2.737510e-01   0   0
#>   3.254948e-01   0.378325759  2.605286e-01   0   0
#>   3.459261e-01   0.366113681  3.679029e-01   0   0
#>  -1.808658e-02   0.976004369  3.131877e-01   0   0
#>  -1.016523e+03   0.178700284 -7.864333e-02   0   0
#>   6.825230e-01   0.269919672  6.733732e-01   0   0
#>   7.815613e-01   0.272970721  8.137293e-01   0   0
#>   8.491136e-01   0.840547492  8.609297e-01   0   0
#>   7.766358e-01   0.807881517  7.685622e-01   0   0
#>   7.377780e-01   0.876011339  6.495072e-01   0   0
#>   7.361614e-01   0.961921677  5.316019e-01   0   0
#>   5.779018e-01   0.157755855  5.512442e-01   0   0
#>   5.972931e-01   0.255232212  5.983985e-01   0   0
#>   6.733732e-01   0.295919947  6.603546e-01   0   0
#>   9.901276e-01   0.621359895  1.003940e+00   0   0
#>   8.620106e-01   0.433169655  8.538323e-01   0   0
#>   8.545223e-01   0.645228754  8.533591e-01   0   0
#>   7.995203e-01   0.249421952  8.290730e-01   0   0
#>   8.556847e-01   0.187703702  9.803946e-01   0   0
#>   1.778883e+01   0.187703702  9.803946e-01   0   0
#>   1.802229e+00   0.394745259  1.266552e+00   0   0
#>   9.600061e-01   0.754570688  9.265477e-01   0   0
#>   5.983985e-01   0.303886368  5.833693e-01   0   0
#>   6.603546e-01   0.358362347  6.526327e-01   0   0
#>   1.733848e-01   0.815623259  9.741819e-02   0   0
#>   1.833418e-01   0.877013372  1.696750e-01   0   0
#>   3.108848e-01   0.885335174  2.834757e-01   0   0
#>   3.327569e-01   1.024659017  3.937027e-01   0   0
#>   3.651504e-01   0.856245288  4.802954e-01   0   0
#>   3.628667e-01   0.701752638  3.619239e-01   0   0
#>   1.739950e-01   0.742635243  1.698969e-01   0   0
#>   1.268691e-01   0.617753074  1.687058e-01   0   0
#>   1.218018e-01   0.478172284  1.962993e-01   0   0
#>   2.946465e-01   0.115452287  2.995071e-01   0   0
#>   2.843212e-01   0.114277591  2.162288e-01   0   0
#>   2.734497e-01   0.156377641  1.764872e-01   0   0
#>   2.183924e-01   0.165259593  1.600723e-01   0   0
#>   2.223727e-01   0.300039067  2.175361e-01   0   0
#>   2.153684e-01  -2.457716122  1.318654e-01   0   0
#>   2.134618e-01   0.027199929  1.060966e-01   0   0
#>   2.483393e-01   0.114277591  2.162288e-01   0   0
#>   4.054124e-01   0.084116069  4.258357e-01   0   0
#>   4.549529e-01   0.004210634  4.558114e-01   0   0
#>   5.018468e-01   0.026857869  5.063426e-01   0   0
#>   7.433939e-01   0.807881517  7.685622e-01   0   0
#>   5.566417e-01   0.704928967  5.192868e-01   0   0
#>   4.912753e-01   1.024659017  3.937027e-01   0   0
#>   9.124481e-01   0.706924264  8.741482e-01   0   0
#>   8.900550e-01   0.779645778  8.451237e-01   0   0
#>   8.932726e-01   0.840547492  8.609297e-01   0   0
#>   8.402593e-01   0.779645778  8.451237e-01   0   0
#>   7.840645e-01   0.687750762  7.867029e-01   0   0
#>   1.962993e-01   0.470876953  2.163587e-01   0   0
#>   3.196563e-02   0.347837059  9.569565e-02   0   0
#>   5.833693e-01   0.310800288  5.196656e-01   0   0
#>   6.526327e-01   0.415566762  6.723393e-01   0   0
#>   8.140096e-01   0.272970721  8.137293e-01   0   0
#>   7.947895e-01   0.418992938  7.663632e-01   0   0
#>   4.558114e-01   0.084116069  4.258357e-01   0   0
#>   5.037786e-01   0.094049626  4.578376e-01   0   0
#>   4.660898e-01   0.267838152  4.727684e-01   0   0
#>   5.094762e-02   0.301412623  9.865057e-02   0   0
#>   1.552022e-01   0.165259593  1.600723e-01   0   0
#>   7.663632e-01   0.422029470  7.357629e-01   0   0
#>   7.746804e-01   0.593054583  6.928396e-01   0   0
#>   8.349944e-01   0.687750762  7.867029e-01   0   0
#>   8.533591e-01   0.706924264  8.741482e-01   0   0
#>   2.162288e-01   0.156377641  1.764872e-01   0   0
#>   5.192868e-01   0.704345298  5.211448e-01   0   0
#>   5.006014e-01   0.600154074  5.510126e-01   0   0
#>   3.619239e-01   0.697425476  3.606253e-01   0   0
#>   6.723393e-01   0.425578612  6.537324e-01   0   0
#>   7.357629e-01   0.593054583  6.928396e-01   0   0
#>   8.137293e-01   0.249421952  8.290730e-01   0   0
#>   8.290730e-01   0.187703702  9.803946e-01   0   0
#>   6.928396e-01   0.608987636  6.667446e-01   0   0
#>   3.606253e-01   0.610632703  4.338164e-01   0   0
#>   3.593037e-01   0.550467081  3.238639e-01   0   0
#>   3.020462e-01   0.563822514  2.222346e-01   0   0
#>   2.406684e-01   0.646440743  1.727691e-01   0   0
#>   1.698969e-01   0.725209820  1.474618e-01   0   0
#>   2.381407e-01   0.460986704  3.512486e-01   0   0
#>   2.235321e-01   0.470876953  2.163587e-01   0   0
#>   1.687058e-01   0.646440743  1.727691e-01   0   0
#>   2.206652e-01   0.550467081  3.238639e-01   0   0
#>   3.238639e-01   0.484272408  3.903930e-01   0   0
#>   4.410217e-01   0.267838152  4.727684e-01   0   0
#>   4.297322e-01   0.366113681  3.679029e-01   0   0
#>   4.429240e-01   0.482566745  4.390800e-01   0   0
#>   4.625622e-01   0.486333958  5.018077e-01   0   0
#>   5.330446e-01   0.423936329  6.160485e-01   0   0
#>   5.323505e-01   0.310800288  5.196656e-01   0   0
#>   6.020842e-01   0.602002897  6.058736e-01   0   0
#>   5.211448e-01   0.592525914  5.761554e-01   0   0
#>   3.512486e-01   0.484272408  3.903930e-01   0   0
#>   7.226353e-01   0.608987636  6.667446e-01   0   0
#>   3.937027e-01   1.038753013  3.915527e-01   0   0
#>   4.727684e-01   0.310800288  5.196656e-01   0   0
#>   6.160485e-01   0.486333958  5.018077e-01   0   0
#>   6.537324e-01   0.595190887  5.991579e-01   0   0
#>   2.834757e-01   0.898594292  2.798966e-01   0   0
#>   5.018077e-01   0.592078640  5.773538e-01   0   0
#>   3.915527e-01   7.972493958  6.646446e-01   0   1
#>   6.045799e-02   0.725209820  1.474618e-01   0   0
#>   6.058736e-01   0.595190887  5.991579e-01   0   0
#>   4.390800e-01   0.484272408  3.903930e-01   0   0
#>   5.773538e-01   0.595190887  5.991579e-01   0   0
#>   8.451237e-01   0.706924264  8.741482e-01   0   0
#>   1.328289e-01   0.725209820  1.474618e-01   0   0
#>   1.896095e-01   0.882282862  1.488643e-01   0   0
#>   1.488643e-01   0.872086439  9.871150e-02   0   0
#>   9.871150e-02   0.896166272  6.692841e-02   0   0
#>   6.692841e-02   1.520110683  6.471541e-02   0   0
#>   6.471541e-02   1.062958390  1.896095e-01   0   0
#>   1.022476e+00   0.778301036  9.898163e-01   0   0
#>   9.898163e-01   0.815540571  9.461517e-01   0   0
#>   9.461517e-01   0.834088159  9.389432e-01   0   0
#>   9.389432e-01   0.869027792  9.472099e-01   0   0
#>   9.472099e-01   0.787429234  1.022476e+00   0   0
#>  -1.585087e+00   0.685918679 -7.377640e-01   0   0
#>  -7.377640e-01   0.702794812 -1.646555e-02   0   0
#>  -1.646555e-02   0.660014388  6.061575e-02   0   0
#>   6.061575e-02   0.636744672  7.103501e-02   0   0
#>   7.103501e-02   0.521173193  4.916716e-02   0   0
#>   4.916716e-02   0.471645337 -5.467543e-02   0   0
#>  -5.467543e-02   0.592572751 -1.585087e+00   0   0
#>   6.754288e-01   0.734990402  6.193844e-01   0   0
#>   6.193844e-01   0.818254213  5.657693e-01   0   0
#>   5.657693e-01   0.863317912  6.404051e-01   0   0
#>   6.404051e-01   0.834391679  6.616044e-01   0   0
#>   6.616044e-01   0.762124031  6.901739e-01   0   0
#>   6.901739e-01   0.738616268  6.754288e-01   0   0
#>   5.807789e-01   0.141260929  5.935437e-01   0   0
#>   5.935437e-01   0.143829961  6.221868e-01   0   0
#>   6.221868e-01   0.128053233  6.745582e-01   0   0
#>   6.745582e-01   0.072609990  7.554305e-01   0   0
#>   7.554305e-01  -2.540278036  1.215583e+00   0   0
#>   1.215583e+00   0.028804487  5.807789e-01   0   0
#>   4.125994e-01   0.230648833  3.942266e-01   0   0
#>   3.942266e-01   0.219838868  3.151658e-01   0   0
#>   3.151658e-01   0.219879183  3.151412e-01   0   0
#>   3.151412e-01   0.296699456  3.077430e-01   0   0
#>   3.077430e-01   0.347347991  3.254948e-01   0   0
#>   3.254948e-01   0.343945149  3.459261e-01   0   0
#>   3.459261e-01   0.249044303  4.125994e-01   0   0
#>  -1.016523e+03 -32.223797793 -1.017604e+03   0   1
#>   6.825230e-01   0.259652269  7.815613e-01   0   0
#>   9.478054e-01   0.873826400  8.491136e-01   0   0
#>   8.491136e-01   0.898722034  7.766358e-01   0   0
#>   7.766358e-01   0.968263257  7.377780e-01   0   0
#>   7.377780e-01   1.366780127  7.361614e-01   0   0
#>   7.361614e-01   3.713377527  1.121575e+00   0   1
#>   5.779018e-01   0.246602795  5.972931e-01   0   0
#>   5.972931e-01   0.269919672  6.733732e-01   0   0
#>   9.901276e-01   0.474334385  8.620106e-01   0   0
#>   8.620106e-01   0.644443099  8.545223e-01   0   0
#>   8.545223e-01   0.621176365  9.901276e-01   0   0
#>   7.995203e-01   0.113909909  8.556847e-01   0   0
#>   8.556847e-01 -15.200275997  1.551029e+01   0   1
#>   1.778883e+01   0.374646611  1.802229e+00   0   0
#>   1.802229e+00   0.621359895  1.003940e+00   0   0
#>   1.003940e+00   0.743340214  9.600061e-01   0   0
#>   5.983985e-01   0.295919947  6.603546e-01   0   0
#>   1.733848e-01   0.801041496  1.833418e-01   0   0
#>   1.833418e-01   0.860696839  3.108848e-01   0   0
#>   3.108848e-01   0.844740504  3.327569e-01   0   0
#>   3.327569e-01   0.802239457  3.651504e-01   0   0
#>   3.651504e-01   0.790952559  3.628667e-01   0   0
#>   3.628667e-01   0.766523781  1.739950e-01   0   0
#>   1.739950e-01   0.767768851  1.733848e-01   0   0
#>   1.268691e-01   0.529819175  1.218018e-01   0   0
#>   2.946465e-01   0.199722032  2.843212e-01   0   0
#>   2.843212e-01   0.200883331  2.734497e-01   0   0
#>   2.734497e-01   0.255014471  2.183924e-01   0   0
#>   2.183924e-01   0.285128470  2.223727e-01   0   0
#>   2.223727e-01   0.263743652  2.737510e-01   0   0
#>   2.153684e-01  -0.238569540  2.134618e-01   0   0
#>   2.134618e-01   0.078991858  2.483393e-01   0   0
#>   2.483393e-01   0.115452287  2.995071e-01   0   0
#>   2.995071e-01   0.112000926  3.820407e-01   0   0
#>   3.820407e-01   0.086459147  4.054124e-01   0   0
#>   4.054124e-01  -0.254818794  2.153684e-01   0   0
#>   1.318654e-01  -0.012738648  4.549529e-01   0   0
#>   4.549529e-01   0.007349032  5.018468e-01   0   0
#>   5.018468e-01 -53.371633739  3.494591e+00   0   1
#>   5.566417e-01   0.856245288  4.802954e-01   0   0
#>   4.802954e-01   0.902610794  4.912753e-01   0   0
#>   4.912753e-01   0.961921677  5.316019e-01   0   0
#>   5.316019e-01   0.876011339  6.495072e-01   0   0
#>   9.265477e-01   0.753055622  9.124481e-01   0   0
#>   9.124481e-01   0.791087393  8.900550e-01   0   0
#>   8.900550e-01   0.804005463  8.932726e-01   0   0
#>   8.402593e-01   0.784682674  7.840645e-01   0   0
#>   7.840645e-01   0.807881517  7.685622e-01   0   0
#>   8.609297e-01   0.781880774  8.402593e-01   0   0
#>   1.962993e-01   0.420436763  3.196563e-02   0   0
#>   5.833693e-01   0.358362347  6.526327e-01   0   0
#>   1.266552e+00   0.284361911  8.140096e-01   0   0
#>   8.140096e-01   0.406407546  7.947895e-01   0   0
#>   7.947895e-01   0.433169655  8.538323e-01   0   0
#>   8.538323e-01   0.394745259  1.266552e+00   0   0
#>   4.558114e-01   0.027942087  5.037786e-01   0   0
#>   5.037786e-01   0.026857869  5.063426e-01   0   0
#>   4.578376e-01   0.084116069  4.258357e-01   0   0
#>   4.660898e-01   0.157755855  5.512442e-01   0   0
#>   5.512442e-01   0.094049626  4.578376e-01   0   0
#>   1.060966e-01   0.178700284 -7.864333e-02   0   0
#>  -7.864333e-02   0.219348887  5.094762e-02   0   0
#>   5.094762e-02   0.165482198  1.552022e-01   0   0
#>   1.552022e-01   0.027199929  1.060966e-01   0   0
#>   7.663632e-01   0.587683846  7.746804e-01   0   0
#>   7.746804e-01   0.647034693  8.349944e-01   0   0
#>   8.349944e-01   0.645228754  8.533591e-01   0   0
#>   5.192868e-01   0.698428760  5.006014e-01   0   0
#>   5.006014e-01   0.701752638  3.619239e-01   0   0
#>   6.723393e-01   0.422029470  7.357629e-01   0   0
#>   3.606253e-01   0.696348124  3.593037e-01   0   0
#>   3.593037e-01   0.676684893  3.020462e-01   0   0
#>   3.020462e-01   0.679121377  2.406684e-01   0   0
#>   2.406684e-01   0.742635243  1.698969e-01   0   0
#>   2.381407e-01   0.378325759  2.605286e-01   0   0
#>   2.605286e-01   0.300039067  2.175361e-01   0   0
#>   2.175361e-01   0.301412623  9.865057e-02   0   0
#>   9.865057e-02   0.347837059  9.569565e-02   0   0
#>   9.569565e-02   0.450682202  2.235321e-01   0   0
#>   2.235321e-01   0.435466773  2.381407e-01   0   0
#>   1.687058e-01   0.563822514  2.222346e-01   0   0
#>   2.222346e-01   0.479789378  2.206652e-01   0   0
#>   2.206652e-01   0.470876953  2.163587e-01   0   0
#>   4.410217e-01   0.403356504  4.297322e-01   0   0
#>   4.297322e-01   0.478919373  4.429240e-01   0   0
#>   4.429240e-01   0.478160187  4.625622e-01   0   0
#>   4.625622e-01   0.372914079  5.330446e-01   0   0
#>   5.330446e-01   0.371950480  5.323505e-01   0   0
#>   5.323505e-01   0.389397303  4.410217e-01   0   0
#>   6.020842e-01   0.704345298  5.211448e-01   0   0
#>   3.512486e-01   0.366113681  3.679029e-01   0   0
#>   1.764872e-01   0.165259593  1.600723e-01   0   0
#>   7.867029e-01   0.687574374  7.226353e-01   0   0
#>   6.160485e-01   0.425578612  6.537324e-01   0   0
#>   1.696750e-01   0.885335174  2.834757e-01   0   0
#>   2.798966e-01   0.976004369  3.131877e-01   0   0
#>   3.131877e-01   1.038753013  3.915527e-01   0   0
#>   9.741819e-02   0.758440713  6.045799e-02   0   0
#>   6.667446e-01   0.602002897  6.058736e-01   0   0
#>   4.390800e-01   0.610632703  4.338164e-01   0   0
#>   4.338164e-01   0.600154074  5.510126e-01   0   0
#>   5.510126e-01   0.592525914  5.761554e-01   0   0
#>   5.761554e-01   0.592078640  5.773538e-01   0   0
#>   1.328289e-01   0.646440743  1.727691e-01   0   0
#>  -1.808658e-02   5.131397124 -5.149098e-01   0   1
#>   9.478054e-01   6.043324333  4.915077e+00   0   1
#>   1.778883e+01  -2.570642583  1.931978e+01   0   1
#>   1.318654e-01  -4.610768687  1.903349e-02   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.948940048 0.13348414
#>   [2,] 0.814724366 0.97964024
#>   [3,] 0.590270586 0.03946092
#>   [4,] 0.795090016 0.65616011
#>   [5,] 0.036476911 0.65495361
#>   [6,] 0.274984116 0.32135059
#>   [7,] 0.948466608 0.00000000
#>   [8,] 0.215220227 0.70879959
#>   [9,] 0.926174804 0.87686932
#>  [10,] 0.255300466 0.63532707
#>  [11,] 0.551103051 0.90171325
#>  [12,] 0.072975889 0.86220625
#>  [13,] 0.741879841 1.00000000
#>  [14,] 0.260710138 0.63366912
#>  [15,] 0.800212368 0.20985315
#>  [16,] 0.581947739 0.08344711
#>  [17,] 0.239531091 0.26328642
#>  [18,] 0.059670771 0.33138096
#>  [19,] 0.000000000 0.49289189
#>  [20,] 0.809436900 0.69245114
#>  [21,] 0.909032880 0.58736357
#>  [22,] 0.792323818 0.92200363
#>  [23,] 0.815471781 0.83884356
#>  [24,] 0.470143752 0.09675557
#>  [25,] 0.285422201 0.61744025
#>  [26,] 0.329266651 0.91455083
#>  [27,] 0.005936734 0.49034881
#>  [28,] 0.157827481 0.43865150
#>  [29,] 0.102683814 0.09400938
#>  [30,] 0.547317362 0.81571581
#>  [31,] 0.163791134 0.25718830
#>  [32,] 0.739999629 0.50744435
#>  [33,] 0.296192583 0.70453352
#>  [34,] 0.179356566 0.75290692
#>  [35,] 0.551549771 0.72987317
#>  [36,] 0.743067006 0.21724436
#>  [37,] 0.349413480 0.17214600
#>  [38,] 0.574206970 0.16889953
#>  [39,] 0.572208237 0.27592023
#>  [40,] 0.782600637 0.95224338
#>  [41,] 0.401239974 0.49030226
#>  [42,] 0.726379911 0.54945389
#>  [43,] 0.417148803 0.34502818
#>  [44,] 0.030617664 0.47813810
#>  [45,] 0.210389579 0.23463513
#>  [46,] 0.024273464 0.02970712
#>  [47,] 0.745852145 0.73465893
#>  [48,] 0.165254442 0.25535210
#>  [49,] 0.951443658 0.40827191
#>  [50,] 0.343683136 0.41913482
#>  [51,] 0.339707884 0.01966239
#>  [52,] 0.272786371 0.29852971
#>  [53,] 0.341348811 0.57345371
#>  [54,] 0.001047047 0.51156716
#>  [55,] 0.078154732 0.16308367
#>  [56,] 0.524078245 0.93268789
#>  [57,] 0.821921516 0.19969919
#>  [58,] 0.835014442 0.89388376
#>  [59,] 0.153506756 0.86794090
#>  [60,] 0.423222070 0.52312648
#>  [61,] 0.422176041 0.11360812
#>  [62,] 0.269423485 0.56564748
#>  [63,] 0.360223063 0.48246668
#>  [64,] 0.975370795 0.33763643
#>  [65,] 0.817811489 0.03413717
#>  [66,] 0.680130470 0.66359757
#>  [67,] 0.552709237 0.49615787
#>  [68,] 0.168377924 0.33592683
#>  [69,] 0.052906485 0.51021096
#>  [70,] 0.009417436 0.42162507
#>  [71,] 0.814911766 0.16073303
#>  [72,] 0.290793858 0.92393505
#>  [73,] 0.807165326 0.86241903
#>  [74,] 0.832325634 0.99872244
#>  [75,] 0.993315194 0.29591132
#>  [76,] 0.925047765 0.59903264
#>  [77,] 0.629346647 0.12673091
#>  [78,] 0.656577786 0.50544486
#>  [79,] 0.994328469 0.46191417
#>  [80,] 0.716714030 0.93012827
#>  [81,] 0.251726532 0.17101736
#>  [82,] 0.714862314 0.10860981
#>  [83,] 1.000000000 0.31791526
#>  [84,] 0.617233743 0.21224922
#>  [85,] 0.758823029 0.12540050
#>  [86,] 0.421888423 0.37202791
#>  [87,] 0.695517016 0.17457066
#>  [88,] 0.935886916 0.19136535
#>  [89,] 0.538981067 0.67979398
#>  [90,] 0.670080345 0.53176755
#>  [91,] 0.547788648 0.37643860
#>  [92,] 0.748597204 0.83550917
#>  [93,] 0.763749913 0.87347464
#>  [94,] 0.497906548 0.32680723
#>  [95,] 0.827747649 0.70072114
#>  [96,] 0.629363692 0.73497964
#>  [97,] 0.105593741 0.89629217
#>  [98,] 0.675019002 0.54180630
#>  [99,] 0.502505158 0.56643043
#> [100,] 0.288230628 0.28355627
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.94894,0.1334841) [5]: 7 65 71 75 88 
#> 2: (0.8147244,0.9796402) [5]: 13 22 40 58 74 
#> 3: (0.5902706,0.03946092) [7]: 7 16 24 51 65 77 82 
#> 4: (0.79509,0.6561601) [6]: 20 21 42 47 66 95 
#> 5: (0.03647691,0.6549536) [6]: 8 10 12 34 54 69 
#> 6: (0.2749841,0.3213506) [7]: 17 28 43 50 52 68 100 
#> 7: (0.9484666,0) [7]: 1 3 46 51 65 75 83 
#> 8: (0.2152202,0.7087996) [4]: 5 10 33 34 
#> 9: (0.9261748,0.8768693) [6]: 23 58 74 76 79 95 
#> 10: (0.2553005,0.6353271) [6]: 5 8 14 33 62 69 
#> 11: (0.5511031,0.9017132) [3]: 30 56 80 
#> 12: (0.07297589,0.8622063) [5]: 5 34 54 59 97 
#> 13: (0.7418798,1) [7]: 2 40 56 72 74 80 97 
#> 14: (0.2607101,0.6336691) [4]: 10 25 33 62 
#> 15: (0.8002124,0.2098531) [7]: 32 36 49 57 64 71 85 
#> 16: (0.5819477,0.08344711) [4]: 3 24 38 77 
#> 17: (0.2395311,0.2632864) [8]: 6 31 45 48 52 68 81 100 
#> 18: (0.05967077,0.331381) [6]: 28 31 46 55 68 70 
#> 19: (0,0.4928919) [4]: 27 46 54 70 
#> 20: (0.8094369,0.6924511) [3]: 4 47 95 
#> 21: (0.9090329,0.5873636) [7]: 4 32 42 49 76 79 95 
#> 22: (0.7923238,0.9220036) [6]: 2 40 58 73 80 93 
#> 23: (0.8154718,0.8388436) [6]: 9 47 58 73 92 95 
#> 24: (0.4701438,0.09675557) [5]: 3 16 38 51 61 
#> 25: (0.2854222,0.6174403) [4]: 14 33 53 62 
#> 26: (0.3292667,0.9145508) [4]: 30 33 56 72 
#> 27: (0.005936734,0.4903488) [5]: 19 44 54 69 70 
#> 28: (0.1578275,0.4386515) [8]: 6 18 44 50 62 68 69 70 
#> 29: (0.1026838,0.09400938) [4]: 46 51 55 81 
#> 30: (0.5473174,0.8157158) [8]: 11 26 33 35 56 80 92 96 
#> 31: (0.1637911,0.2571883) [5]: 17 18 48 55 68 
#> 32: (0.7399996,0.5074444) [7]: 15 21 36 42 49 78 90 
#> 33: (0.2961926,0.7045335) [11]: 8 10 14 25 26 30 34 35 53 72 89 
#> 34: (0.1793566,0.7529069) [6]: 5 8 12 33 59 72 
#> 35: (0.5515498,0.7298732) [4]: 30 33 89 96 
#> 36: (0.743067,0.2172444) [8]: 15 32 39 78 84 85 87 91 
#> 37: (0.3494135,0.172146) [6]: 43 51 61 81 94 100 
#> 38: (0.574207,0.1688995) [7]: 16 24 39 61 77 84 94 
#> 39: (0.5722082,0.2759202) [5]: 36 38 84 91 94 
#> 40: (0.7826006,0.9522434) [4]: 2 13 22 80 
#> 41: (0.40124,0.4903023) [6]: 50 53 60 63 67 86 
#> 42: (0.7263799,0.5494539) [6]: 4 21 32 66 90 98 
#> 43: (0.4171488,0.3450282) [6]: 6 37 50 86 94 100 
#> 44: (0.03061766,0.4781381) [4]: 27 28 69 70 
#> 45: (0.2103896,0.2346351) [4]: 17 48 55 81 
#> 46: (0.02427346,0.02970712) [7]: 7 18 19 29 51 55 70 
#> 47: (0.7458521,0.7346589) [7]: 4 20 23 66 92 95 96 
#> 48: (0.1652544,0.2553521) [4]: 17 31 45 55 
#> 49: (0.9514437,0.4082719) [5]: 15 21 32 64 79 
#> 50: (0.3436831,0.4191348) [7]: 6 28 41 43 62 63 86 
#> 51: (0.3397079,0.01966239) [8]: 3 7 24 29 37 46 61 81 
#> 52: (0.2727864,0.2985297) [3]: 6 17 100 
#> 53: (0.3413488,0.5734537) [8]: 25 33 41 60 62 63 89 99 
#> 54: (0.001047047,0.5115672) [5]: 5 12 19 27 69 
#> 55: (0.07815473,0.1630837) [7]: 18 29 31 45 46 48 81 
#> 56: (0.5240782,0.9326879) [6]: 11 13 26 30 72 80 
#> 57: (0.8219215,0.1996992) [4]: 15 64 71 88 
#> 58: (0.8350144,0.8938838) [6]: 2 9 22 23 73 74 
#> 59: (0.1535068,0.8679409) [4]: 12 34 72 97 
#> 60: (0.4232221,0.5231265) [4]: 41 53 67 99 
#> 61: (0.422176,0.1136081) [5]: 24 37 38 51 94 
#> 62: (0.2694235,0.5656475) [8]: 10 14 25 28 50 53 63 69 
#> 63: (0.3602231,0.4824667) [4]: 41 50 53 62 
#> 64: (0.9753708,0.3376364) [7]: 15 49 57 75 79 83 88 
#> 65: (0.8178115,0.03413717) [6]: 1 3 7 71 82 85 
#> 66: (0.6801305,0.6635976) [6]: 4 42 47 89 96 98 
#> 67: (0.5527092,0.4961579) [8]: 41 60 78 86 90 91 98 99 
#> 68: (0.1683779,0.3359268) [5]: 6 17 18 28 31 
#> 69: (0.05290649,0.510211) [7]: 5 10 27 28 44 54 62 
#> 70: (0.009417436,0.4216251) [6]: 18 19 27 28 44 46 
#> 71: (0.8149118,0.160733) [6]: 1 15 57 65 85 88 
#> 72: (0.2907939,0.923935) [7]: 13 26 33 34 56 59 97 
#> 73: (0.8071653,0.862419) [5]: 22 23 58 92 93 
#> 74: (0.8323256,0.9987224) [4]: 2 9 13 58 
#> 75: (0.9933152,0.2959113) [5]: 1 7 64 83 88 
#> 76: (0.9250478,0.5990326) [4]: 9 21 79 95 
#> 77: (0.6293466,0.1267309) [6]: 3 16 38 82 84 87 
#> 78: (0.6565778,0.5054449) [5]: 32 36 67 90 91 
#> 79: (0.9943285,0.4619142) [6]: 9 21 49 64 76 83 
#> 80: (0.716714,0.9301283) [8]: 11 13 22 30 40 56 92 93 
#> 81: (0.2517265,0.1710174) [7]: 17 29 37 45 51 55 100 
#> 82: (0.7148623,0.1086098) [5]: 3 65 77 85 87 
#> 83: (1,0.3179153) [4]: 7 64 75 79 
#> 84: (0.6172337,0.2122492) [5]: 36 38 39 77 87 
#> 85: (0.758823,0.1254005) [6]: 15 36 65 71 82 87 
#> 86: (0.4218884,0.3720279) [6]: 41 43 50 67 91 94 
#> 87: (0.695517,0.1745707) [5]: 36 77 82 84 85 
#> 88: (0.9358869,0.1913653) [5]: 1 57 64 71 75 
#> 89: (0.5389811,0.679794) [7]: 33 35 53 66 96 98 99 
#> 90: (0.6700803,0.5317676) [5]: 32 42 67 78 98 
#> 91: (0.5477886,0.3764386) [6]: 36 39 67 78 86 94 
#> 92: (0.7485972,0.8355092) [7]: 23 30 47 73 80 93 96 
#> 93: (0.7637499,0.8734746) [4]: 22 73 80 92 
#> 94: (0.4979065,0.3268072) [7]: 37 38 39 43 61 86 91 
#> 95: (0.8277476,0.7007211) [7]: 4 9 20 21 23 47 76 
#> 96: (0.6293637,0.7349796) [6]: 30 35 47 66 89 92 
#> 97: (0.1055937,0.8962922) [4]: 12 13 59 72 
#> 98: (0.675019,0.5418063) [6]: 42 66 67 89 90 99 
#> 99: (0.5025052,0.5664304) [5]: 53 60 67 89 98 
#> 100: (0.2882306,0.2835563) [6]: 6 17 37 43 52 81 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  7 9 12 13 19 46 54 74 79 83 97 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.119280623 0.124793130 0.126629065 0.135041130 0.299290666 0.090454816 
#>                                                                         
#> 0.062760482 0.031764485 0.051119629 0.104873151 0.609076180 0.127061823 
#>                                                                         
#> 0.087414549 0.064188348 0.112593279 0.151515261 0.096793886 0.123157582 
#>                                                                         
#> 0.186545575 0.139489166 0.020091702 0.076318534 0.051877169 0.237944036 
#>                                                                         
#> 0.083693709 0.056847678 0.152522764 0.357940686 0.186877567 0.103265043 
#>                                                                         
#> 0.101682849 0.048456194 0.021511137 0.142823987 0.104350367 0.022996962 
#>                                                                         
#> 0.315594312 0.080384692 0.126454103 0.058406671 0.138666866 0.140740415 
#>                                                                         
#> 0.153627112 0.223501273 0.068577447 0.079251291 0.192652789 0.119258219 
#>                                                                         
#> 0.204414874 0.117902630 0.055672720 0.187251733 0.233448533 0.069499089 
#>                                                                         
#> 0.054207473 0.087756704 0.066289778 0.039591638 0.206333140 0.074578005 
#>                                                                         
#> 0.233546034 0.290216198 0.093185558 0.069415964 0.126625053 0.078872011 
#>                                                                         
#> 0.002347961 0.049662595 0.075868073 0.118311261 0.392199421 0.143848865 
#>                                                                         
#> 0.127237085 0.149383535 0.068811827 0.021774463 0.088649010 0.044162178 
#>                                                                         
#> 0.068677502 0.073634280 0.044800944 0.042019073 0.064454369 0.100887589 
#>                                                                         
#> 0.161773124 0.125007187 0.072347701 0.138639827 0.219467936 0.274636407 
#>                                                                         
#> 0.060358696 0.039057133 0.164257281 0.175068690 0.173753023 0.256611361 
#>                                                                         
#> 0.077981295 0.155957738 0.099846433 0.126885077 0.059012266 0.029583834 
#>                                                                         
#> 0.300898510 0.224425404 0.051632323 0.204129594 0.148266964 0.105898423 
#>                                                                         
#> 0.168753351 0.103442050 0.077982649 0.086878909 0.080158321 0.082787778 
#>                                                                         
#> 0.226249801 0.086371892 0.090056803 0.070366868 0.065456042 0.091296923 
#>                                                                         
#> 0.180339495 0.132265763 0.096104426 0.092924040 0.121898481 0.011187794 
#>                                                                         
#> 0.088451483 0.087593759 0.125715355 0.123140605 0.090338475 0.119087310 
#>                                                                         
#> 0.151506460 0.086363659 0.144110558 0.047058172 0.193770278 0.125977461 
#>                                                                         
#> 0.174262371 0.040877644 0.068739172 0.059334795 0.136770181 0.164512976 
#>                                                                         
#> 0.133484983 0.168379749 0.075636265 0.042219926 0.061836587 0.088124122 
#>                                                                         
#> 0.025960256 0.360363087 0.227603174 0.142494448 0.095618984 0.044766669 
#>                                                                         
#> 0.133090694 0.251343688 0.115199883 0.126914520 0.133101236 0.055246783 
#>                                                                         
#> 0.039023997 0.092663011 0.145672122 0.219701956 0.186677695 0.173232346 
#>                                                                         
#> 0.210441993 0.147698853 0.165786552 0.107598081 0.068032080 0.022926460 
#>                                                                         
#> 0.040048478 0.144122956 0.119504442 0.924670471 0.081084659 0.092734582 
#>                                                                         
#> 0.117051773 0.201782273 0.277838969 0.420514801 0.071096446 0.005658032 
#>                                                                         
#> 0.041106793 0.086080720 0.168030978 0.080734792 0.047178113 0.457454305 
#>                                                                         
#> 0.227965909 0.074265587 0.029564545 0.051272390 0.023966434 0.216815687 
#>                                                                         
#> 0.249481279 0.303621636 0.057621372 0.094049603 0.085802303 0.075985053 
#>                                                                         
#> 0.074699225 0.040867153 0.093071525 0.052749516 0.303743437 0.169309284 
#>                                                                         
#> 0.127849909 0.108802160 0.145401888 0.103292778 0.071886357 0.006458490 
#>                                                                         
#> 0.018704590 0.186974123 0.184044825 0.151699998 0.019815237 0.076045050 
#>                                                                         
#> 0.056316346 0.061405169 0.066957652 0.125304953 0.024995999 0.050842008 
#>                                                                         
#> 0.071151959 0.039600750 0.212605649 0.239404412 0.195654058 0.027536335 
#>                                                                         
#> 0.050996691 0.133197332 0.169061075 0.127000675 0.101404950 0.248410718 
#>                                                                         
#> 0.167761621 0.073300296 0.085946921 0.115108337 0.202250716 0.074029213 
#>                                                                         
#> 0.083445801 0.244045677 0.251945313 0.180653194 0.125932369 0.063890894 
#>                                                                         
#> 0.185678005 0.127104641 0.097693469 0.152792178 0.093386689 0.214406829 
#>                                                                         
#> 0.061077812 0.107039367 0.175375590 0.091529200 0.120063240 0.151582406 
#>                                                                         
#> 0.039504958 0.102474945 0.041758632 0.051927147 0.027412580 0.150351811 
#>                                                                         
#> 0.116488895 0.161309313 0.114269705 0.045420115 0.031551899 0.108667531 
#>                                                                         
#> 0.142075607 0.119820356 0.104282902 0.122654093 0.130550593 0.081652673 
#>                                                 
#> 0.322064907 0.153804580 0.644682335 0.463820376 
#> 
#> $alpha
#> [1] 0.2967925
#> 
#> attr(,"class")
#> [1] "scree"