Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, alpha = c("rahman", "q90", "omega"), ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

alpha

character, numeric, or function. Controls the alpha radius. Valid character values are: - "rahman" (default): Rahman's MST-based middle-50% alpha - "q90": 90th percentile of MST edge lengths - "omega": graph-theoretic scagnostics alpha Alternatively: - a numeric value giving a fixed alpha - a function with no arguments that returns a single numeric alpha

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    97   95 0.060067275 0.97954186 0.033622129 0.85798781   0.01922427
#>    97    1 0.060067275 0.97954186 0.085780755 0.90934941   0.10870757
#>    87    1 0.155328050 0.98355162 0.085780755 0.90934941   0.14716410
#>     1   95 0.085780755 0.90934941 0.033622129 0.85798781   0.11313392
#>    32   85 0.008711212 0.74973864 0.109585001 0.71414906   0.07976456
#>    97   87 0.060067275 0.97954186 0.155328050 0.98355162   0.10076143
#>    87   86 0.155328050 0.98355162 0.185894234 0.87235958   0.19308904
#>    85   86 0.109585001 0.71414906 0.185894234 0.87235958   0.11505833
#>    22   86 0.182670854 0.99544852 0.185894234 0.87235958   0.23717256
#>    86    9 0.185894234 0.87235958 0.165582466 0.72701619   0.19367789
#>    85    9 0.109585001 0.71414906 0.165582466 0.72701619   0.11753796
#>    39   86 0.253545513 0.85578112 0.185894234 0.87235958   0.20888982
#>    18    9 0.266174387 0.81855108 0.165582466 0.72701619   0.22088019
#>    82   26 0.240525224 0.65567552 0.271023169 0.72075465   0.21725647
#>    22   39 0.182670854 0.99544852 0.253545513 0.85578112   0.26085615
#>    39   18 0.253545513 0.85578112 0.266174387 0.81855108   0.31980527
#>     6   16 0.383225659 0.81834060 0.347738310 0.91760896   0.32476291
#>    85   13 0.109585001 0.71414906 0.195516273 0.56660089   0.15560513
#>    13   82 0.195516273 0.56660089 0.240525224 0.65567552   0.15636044
#>    26   36 0.271023169 0.72075465 0.378788670 0.73716598   0.31833344
#>    36    6 0.378788670 0.73716598 0.383225659 0.81834060   0.32463239
#>    32   15 0.008711212 0.74973864 0.054522262 0.46175393   0.01361356
#>    33   36 0.370835275 0.66654812 0.378788670 0.73716598   0.32823453
#>    27    6 0.441586369 0.86631570 0.383225659 0.81834060   0.43854596
#>    33    7 0.370835275 0.66654812 0.456282136 0.69871753   0.40771629
#>    82   33 0.240525224 0.65567552 0.370835275 0.66654812   0.30535760
#>    30   56 0.461153054 0.94359077 0.464599833 1.00000000   0.39404484
#>    16   27 0.347738310 0.91760896 0.441586369 0.86631570   0.40779326
#>    56   22 0.464599833 1.00000000 0.182670854 0.99544852   0.32234422
#>    72   30 0.495857469 0.98379445 0.461153054 0.94359077   0.50831366
#>    30   27 0.461153054 0.94359077 0.441586369 0.86631570   0.47471830
#>     7   41 0.456282136 0.69871753 0.475447019 0.85247777   0.45078056
#>    20   73 0.328980281 0.57967761 0.352568135 0.58819035   0.32173618
#>    55  100 0.016717392 0.38071851 0.163434464 0.36855882   0.09176695
#>    15   13 0.054522262 0.46175393 0.195516273 0.56660089   0.06785549
#>    36    7 0.378788670 0.73716598 0.456282136 0.69871753   0.44546293
#>    15  100 0.054522262 0.46175393 0.163434464 0.36855882   0.15695793
#>    72   56 0.495857469 0.98379445 0.464599833 1.00000000   0.46959403
#>    37   21 0.523988381 0.66300066 0.593333424 0.77225285   0.52185505
#>    73    7 0.352568135 0.58819035 0.456282136 0.69871753   0.44135565
#>    30   41 0.461153054 0.94359077 0.475447019 0.85247777   0.49171808
#>    20   13 0.328980281 0.57967761 0.195516273 0.56660089   0.26951234
#>    72   53 0.495857469 0.98379445 0.543394362 0.90594353   0.56418113
#>    75   19 0.581373540 0.83562665 0.566732253 0.90280052   0.54961092
#>    20   94 0.328980281 0.57967761 0.236648898 0.40431577   0.29494070
#>    21   41 0.593333424 0.77225285 0.475447019 0.85247777   0.50596724
#>    37    7 0.523988381 0.66300066 0.456282136 0.69871753   0.44855724
#>    41   75 0.475447019 0.85247777 0.581373540 0.83562665   0.52986927
#>   100   94 0.163434464 0.36855882 0.236648898 0.40431577   0.15877501
#>    37   73 0.523988381 0.66300066 0.352568135 0.58819035   0.46214881
#>    43   31 0.394931773 0.48264437 0.384086693 0.43575161   0.43602462
#>    37   43 0.523988381 0.66300066 0.394931773 0.48264437   0.47132070
#>    43   94 0.394931773 0.48264437 0.236648898 0.40431577   0.29747933
#>   100   48 0.163434464 0.36855882 0.139567148 0.31890296   0.21098513
#>    76   62 0.631598421 0.76201424 0.667002992 0.85402435   0.63054563
#>    75   21 0.581373540 0.83562665 0.593333424 0.77225285   0.62417346
#>    37   99 0.523988381 0.66300066 0.609150991 0.67220759   0.56134599
#>    37   92 0.523988381 0.66300066 0.573855647 0.61539915   0.56753270
#>    48   55 0.139567148 0.31890296 0.016717392 0.38071851   0.06468681
#>    55   32 0.016717392 0.38071851 0.008711212 0.74973864  -1.17296918
#>    99   76 0.609150991 0.67220759 0.631598421 0.76201424   0.60041615
#>    72   19 0.495857469 0.98379445 0.566732253 0.90280052   0.63330305
#>    62   19 0.667002992 0.85402435 0.566732253 0.90280052   0.61694004
#>    74   51 0.774595721 0.75833534 0.717124235 0.82752652   0.70302442
#>    45   99 0.669941403 0.68539284 0.609150991 0.67220759   0.64884882
#>     8   83 0.308203552 0.21655297 0.397124396 0.32640190   0.29425953
#>    99   92 0.609150991 0.67220759 0.573855647 0.61539915   0.63193022
#>    51   84 0.717124235 0.82752652 0.769235633 0.97972355   0.72803251
#>    31   94 0.384086693 0.43575161 0.236648898 0.40431577   0.32045724
#>    48   69 0.139567148 0.31890296 0.000000000 0.28512195   0.08351625
#>    31   83 0.384086693 0.43575161 0.397124396 0.32640190   0.44345342
#>    11   31 0.489207150 0.45393267 0.384086693 0.43575161   0.44587587
#>    45   93 0.669941403 0.68539284 0.685851605 0.59686983   0.73090103
#>    34   93 0.556451999 0.40544079 0.685851605 0.59686983   0.61335695
#>    11   34 0.489207150 0.45393267 0.556451999 0.40544079   0.58015177
#>    50   58 0.846843878 0.88752615 0.814306253 0.94741108   0.76958599
#>    44    8 0.204642149 0.24390424 0.308203552 0.21655297   0.27818539
#>    44   46 0.204642149 0.24390424 0.047823732 0.15990143   0.15030529
#>    44   48 0.204642149 0.24390424 0.139567148 0.31890296   0.11171524
#>    83   28 0.397124396 0.32640190 0.365295705 0.23509557   0.41302300
#>    45   74 0.669941403 0.68539284 0.774595721 0.75833534   0.70282900
#>    63   25 0.503425305 0.20966861 0.540918712 0.28865955   0.46128712
#>    49   83 0.514491122 0.35877566 0.397124396 0.32640190   0.46729374
#>    28    8 0.365295705 0.23509557 0.308203552 0.21655297   0.35167382
#>    59    8 0.398169243 0.21263363 0.308203552 0.21655297   0.35106283
#>    84   58 0.769235633 0.97972355 0.814306253 0.94741108   0.79672806
#>    66   17 0.836121750 0.94485523 0.825617362 0.97306127   0.82652709
#>    66   58 0.836121750 0.94485523 0.814306253 0.94741108   0.82126298
#>    74    3 0.774595721 0.75833534 0.769280914 0.59192468   0.77498611
#>    93   12 0.685851605 0.59686983 0.779055593 0.32990144   0.71798832
#>    49   25 0.514491122 0.35877566 0.540918712 0.28865955   0.57893457
#>    25   88 0.540918712 0.28865955 0.661256009 0.21510864   0.63581117
#>    74   54 0.774595721 0.75833534 0.835360153 0.77197338   0.79043401
#>    54   50 0.835360153 0.77197338 0.846843878 0.88752615   0.79260226
#>     8   40 0.308203552 0.21655297 0.282643632 0.07283984   0.33131246
#>    63   10 0.503425305 0.20966861 0.527698348 0.21362291   0.50838354
#>    59   63 0.398169243 0.21263363 0.503425305 0.20966861   0.45245562
#>    68   70 0.875340411 0.78040643 0.889589371 0.87510046   0.85389690
#>    89   68 0.857451691 0.66324642 0.875340411 0.78040643   0.86683121
#>    70    5 0.889589371 0.87510046 0.973063603 0.90619041   0.91051892
#>    70   66 0.889589371 0.87510046 0.836121750 0.94485523   0.88047503
#>    25   10 0.540918712 0.28865955 0.527698348 0.21362291   0.54052475
#>    63   61 0.503425305 0.20966861 0.412830924 0.14150496   0.48808228
#>    79   40 0.245048732 0.07672028 0.282643632 0.07283984   0.26677329
#>    60   79 0.209977329 0.02773072 0.245048732 0.07672028   0.20651409
#>    25   24 0.540918712 0.28865955 0.538764687 0.21147062   0.59896711
#>    34   25 0.556451999 0.40544079 0.540918712 0.28865955   0.64183657
#>    52   40 0.233194915 0.09662820 0.282643632 0.07283984   0.28744787
#>    23   68 0.949973906 0.70991366 0.875340411 0.78040643   0.88756520
#>    81   80 0.868384436 0.60041588 0.920589746 0.65465738   0.88587472
#>    89    3 0.857451691 0.66324642 0.769280914 0.59192468   0.81646973
#>    40   65 0.282643632 0.07283984 0.507626884 0.03519486   0.39047211
#>    40   60 0.282643632 0.07283984 0.209977329 0.02773072   0.36304579
#>    46   60 0.047823732 0.15990143 0.209977329 0.02773072   0.12864300
#>    46   69 0.047823732 0.15990143 0.000000000 0.28512195  -0.22085007
#>     3   93 0.769280914 0.59192468 0.685851605 0.59686983   0.71954561
#>    47   17 1.000000000 0.95294441 0.825617362 0.97306127   0.91599946
#>    80   23 0.920589746 0.65465738 0.949973906 0.70991366   0.89510161
#>    23    5 0.949973906 0.70991366 0.973063603 0.90619041   0.97100574
#>    24   10 0.538764687 0.21147062 0.527698348 0.21362291   0.52380529
#>    64   88 0.658367104 0.20681622 0.661256009 0.21510864   0.59945496
#>    78   64 0.563535575 0.13422872 0.658367104 0.20681622   0.59773488
#>     3   77 0.769280914 0.59192468 0.859573416 0.41478276   0.81294765
#>    81    3 0.868384436 0.60041588 0.769280914 0.59192468   0.82493397
#>    14   12 0.712180549 0.24891512 0.779055593 0.32990144   0.66405190
#>    14   88 0.712180549 0.24891512 0.661256009 0.21510864   0.65432050
#>    91   38 0.548806580 0.04176966 0.597772123 0.08112133   0.55132933
#>    96   23 0.911510248 0.51568277 0.949973906 0.70991366   1.14617778
#>    35   12 0.857967708 0.34844030 0.779055593 0.32990144   0.84780621
#>    84   17 0.769235633 0.97972355 0.825617362 0.97306127   5.72618897
#>     5   47 0.973063603 0.90619041 1.000000000 0.95294441   0.91703888
#>    96   90 0.911510248 0.51568277 0.928446194 0.38694202   1.15310560
#>    57   77 0.805812576 0.34648104 0.859573416 0.41478276   0.73088120
#>    65   78 0.507626884 0.03519486 0.563535575 0.13422872   0.48694258
#>    60   29 0.209977329 0.02773072 0.016823875 0.00000000   0.41127232
#>    35   90 0.857967708 0.34844030 0.928446194 0.38694202   0.88596191
#>    35   57 0.857967708 0.34844030 0.805812576 0.34648104   0.83480726
#>    64   38 0.658367104 0.20681622 0.597772123 0.08112133   0.69685834
#>    67   64 0.707496939 0.24160196 0.658367104 0.20681622   0.74309860
#>    65   91 0.507626884 0.03519486 0.548806580 0.04176966   0.51937491
#>    71   47 0.999267224 0.29267144 1.000000000 0.95294441   1.99384526
#>    14   67 0.712180549 0.24891512 0.707496939 0.24160196   0.81747852
#>    67    2 0.707496939 0.24160196 0.765146698 0.03311526   0.80439113
#>    98    2 0.999819740 0.01258032 0.765146698 0.03311526   0.61126578
#>    71   35 0.999267224 0.29267144 0.857967708 0.34844030   0.90928575
#>     4   35 0.993550233 0.23361283 0.857967708 0.34844030   0.86462151
#>    71    4 0.999267224 0.29267144 0.993550233 0.23361283   2.11534440
#>     4   42 0.993550233 0.23361283 0.921809657 0.10234640   1.03767667
#>    42    2 0.921809657 0.10234640 0.765146698 0.03311526   0.87931459
#>    29   98 0.016823875 0.00000000 0.999819740 0.01258032   0.56160829
#>    86    1 0.185894234 0.87235958 0.085780755 0.90934941   0.14716410
#>    86   95 0.185894234 0.87235958 0.033622129 0.85798781   0.11313392
#>    95   85 0.033622129 0.85798781 0.109585001 0.71414906   0.07976456
#>    32   97 0.008711212 0.74973864 0.060067275 0.97954186  -9.56650505
#>    22   87 0.182670854 0.99544852 0.155328050 0.98355162   0.10076143
#>    18   86 0.266174387 0.81855108 0.185894234 0.87235958   0.20888982
#>    26    9 0.271023169 0.72075465 0.165582466 0.72701619   0.22088019
#>    16   39 0.347738310 0.91760896 0.253545513 0.85578112   0.26085615
#>    16   18 0.347738310 0.91760896 0.266174387 0.81855108   0.31980527
#>     9   13 0.165582466 0.72701619 0.195516273 0.56660089   0.15560513
#>     9   82 0.165582466 0.72701619 0.240525224 0.65567552   0.15636044
#>    18   36 0.266174387 0.81855108 0.378788670 0.73716598   0.31833344
#>    18    6 0.266174387 0.81855108 0.383225659 0.81834060   0.32463239
#>    85   15 0.109585001 0.71414906 0.054522262 0.46175393   0.01361356
#>    41    6 0.475447019 0.85247777 0.383225659 0.81834060   0.43854596
#>    26   33 0.271023169 0.72075465 0.370835275 0.66654812   0.30535760
#>    16   56 0.347738310 0.91760896 0.464599833 1.00000000   0.39404484
#>    16   22 0.347738310 0.91760896 0.182670854 0.99544852   0.32234422
#>    53   30 0.543394362 0.90594353 0.461153054 0.94359077   0.50831366
#>    41   27 0.475447019 0.85247777 0.441586369 0.86631570   0.47471830
#>    33   73 0.370835275 0.66654812 0.352568135 0.58819035   0.32173618
#>    20   82 0.328980281 0.57967761 0.240525224 0.65567552   0.30713111
#>     6    7 0.383225659 0.81834060 0.456282136 0.69871753   0.44546293
#>    13  100 0.195516273 0.56660089 0.163434464 0.36855882   0.15695793
#>     7   21 0.456282136 0.69871753 0.593333424 0.77225285   0.52185505
#>    53   41 0.543394362 0.90594353 0.475447019 0.85247777   0.49171808
#>    94   13 0.236648898 0.40431577 0.195516273 0.56660089   0.26951234
#>    19   53 0.566732253 0.90280052 0.543394362 0.90594353   0.56418113
#>    43   20 0.394931773 0.48264437 0.328980281 0.57967761   0.36024053
#>    53   75 0.543394362 0.90594353 0.581373540 0.83562665   0.52986927
#>    43   73 0.394931773 0.48264437 0.352568135 0.58819035   0.46214881
#>    11   43 0.489207150 0.45393267 0.394931773 0.48264437   0.47132070
#>    44  100 0.204642149 0.24390424 0.163434464 0.36855882   0.23154347
#>    75   62 0.581373540 0.83562665 0.667002992 0.85402435   0.63054563
#>    76   21 0.631598421 0.76201424 0.593333424 0.77225285   0.62417346
#>    21   99 0.593333424 0.77225285 0.609150991 0.67220759   0.56134599
#>    69   55 0.000000000 0.28512195 0.016717392 0.38071851   0.06468681
#>    69   32 0.000000000 0.28512195 0.008711212 0.74973864  -1.17296918
#>    84   19 0.769235633 0.97972355 0.566732253 0.90280052   0.63330305
#>    76   51 0.631598421 0.76201424 0.717124235 0.82752652   0.70302442
#>    93   99 0.685851605 0.59686983 0.609150991 0.67220759   0.64884882
#>    11   37 0.489207150 0.45393267 0.523988381 0.66300066   0.47661789
#>    94   83 0.236648898 0.40431577 0.397124396 0.32640190   0.29425953
#>    76   45 0.631598421 0.76201424 0.669941403 0.68539284   0.63185935
#>    93   92 0.685851605 0.59686983 0.573855647 0.61539915   0.63193022
#>    62   84 0.667002992 0.85402435 0.769235633 0.97972355   0.72803251
#>    49   31 0.514491122 0.35877566 0.384086693 0.43575161   0.44587587
#>    92   34 0.573855647 0.61539915 0.556451999 0.40544079   0.58015177
#>    84   72 0.769235633 0.97972355 0.495857469 0.98379445   0.63720628
#>    51   58 0.717124235 0.82752652 0.814306253 0.94741108   0.76958599
#>    49   11 0.514491122 0.35877566 0.489207150 0.45393267   0.50698868
#>    94    8 0.236648898 0.40431577 0.308203552 0.21655297   0.27818539
#>    52   46 0.233194915 0.09662820 0.047823732 0.15990143   0.15030529
#>    46   48 0.047823732 0.15990143 0.139567148 0.31890296   0.11171524
#>    59   28 0.398169243 0.21263363 0.365295705 0.23509557   0.41302300
#>    76   74 0.631598421 0.76201424 0.774595721 0.75833534   0.70282900
#>    83   25 0.397124396 0.32640190 0.540918712 0.28865955   0.46128712
#>    61    8 0.412830924 0.14150496 0.308203552 0.21655297   0.35106283
#>    17   58 0.825617362 0.97306127 0.814306253 0.94741108   0.79672806
#>     3   45 0.769280914 0.59192468 0.669941403 0.68539284   0.75445227
#>    57   12 0.805812576 0.34648104 0.779055593 0.32990144   0.71798832
#>    51   54 0.717124235 0.82752652 0.835360153 0.77197338   0.79043401
#>    51   50 0.717124235 0.82752652 0.846843878 0.88752615   0.79260226
#>    61   40 0.412830924 0.14150496 0.282643632 0.07283984   0.33131246
#>    52   44 0.233194915 0.09662820 0.204642149 0.24390424   0.24175571
#>    83   63 0.397124396 0.32640190 0.503425305 0.20966861   0.45245562
#>    50   70 0.846843878 0.88752615 0.889589371 0.87510046   0.85389690
#>    68   54 0.875340411 0.78040643 0.835360153 0.77197338   0.84411606
#>    66    5 0.836121750 0.94485523 0.973063603 0.90619041   0.91051892
#>    50   66 0.846843878 0.88752615 0.836121750 0.94485523   0.88047503
#>    78   61 0.563535575 0.13422872 0.412830924 0.14150496   0.48808228
#>    61   59 0.412830924 0.14150496 0.398169243 0.21263363   0.45009615
#>    52   79 0.233194915 0.09662820 0.245048732 0.07672028   0.20651409
#>    89   74 0.857451691 0.66324642 0.774595721 0.75833534   0.81699031
#>    88   24 0.661256009 0.21510864 0.538764687 0.21147062   0.59896711
#>    14   25 0.712180549 0.24891512 0.540918712 0.28865955   0.64183657
#>    78   63 0.563535575 0.13422872 0.503425305 0.20966861   0.52334455
#>    89   80 0.857451691 0.66324642 0.920589746 0.65465738   0.88587472
#>    61   65 0.412830924 0.14150496 0.507626884 0.03519486   0.39047211
#>    65   60 0.507626884 0.03519486 0.209977329 0.02773072   0.36304579
#>    52   60 0.233194915 0.09662820 0.209977329 0.02773072   0.12864300
#>    29   46 0.016823875 0.00000000 0.047823732 0.15990143   0.10596245
#>    29   69 0.016823875 0.00000000 0.000000000 0.28512195  -0.22085007
#>    57   93 0.805812576 0.34648104 0.685851605 0.59686983   0.71954561
#>    89   23 0.857451691 0.66324642 0.949973906 0.70991366   0.89510161
#>    68    5 0.875340411 0.78040643 0.973063603 0.90619041   0.97100574
#>    12   34 0.779055593 0.32990144 0.556451999 0.40544079   0.69590389
#>    78   10 0.563535575 0.13422872 0.527698348 0.21362291   0.52380529
#>    24   64 0.538764687 0.21147062 0.658367104 0.20681622   0.59773488
#>    96   77 0.911510248 0.51568277 0.859573416 0.41478276   0.81294765
#>    96    3 0.911510248 0.51568277 0.769280914 0.59192468   0.82493397
#>    42   14 0.921809657 0.10234640 0.712180549 0.24891512   0.84130245
#>    67   88 0.707496939 0.24160196 0.661256009 0.21510864   0.65432050
#>    78   38 0.563535575 0.13422872 0.597772123 0.08112133   0.55132933
#>    42   12 0.921809657 0.10234640 0.779055593 0.32990144   0.84780621
#>    66   47 0.836121750 0.94485523 1.000000000 0.95294441   0.91703888
#>    47   23 1.000000000 0.95294441 0.949973906 0.70991366   1.25798239
#>    71   90 0.999267224 0.29267144 0.928446194 0.38694202   1.15310560
#>    96   81 0.911510248 0.51568277 0.868384436 0.60041588   0.94014143
#>    65   29 0.507626884 0.03519486 0.016823875 0.00000000   0.41127232
#>    77   90 0.859573416 0.41478276 0.928446194 0.38694202   0.88596191
#>     2   38 0.765146698 0.03311526 0.597772123 0.08112133   0.69685834
#>     2   64 0.765146698 0.03311526 0.658367104 0.20681622   0.74309860
#>    78   91 0.563535575 0.13422872 0.548806580 0.04176966   0.51937491
#>    42   67 0.921809657 0.10234640 0.707496939 0.24160196   0.81747852
#>     2   91 0.765146698 0.03311526 0.548806580 0.04176966   0.65392316
#>     2   65 0.765146698 0.03311526 0.507626884 0.03519486   0.63121233
#>    71   96 0.999267224 0.29267144 0.911510248 0.51568277   1.29981251
#>    42   35 0.921809657 0.10234640 0.857967708 0.34844030   0.86462151
#>    98   71 0.999819740 0.01258032 0.999267224 0.29267144 153.53529033
#>    98    4 0.999819740 0.01258032 0.993550233 0.23361283   2.11534440
#>    98   42 0.999819740 0.01258032 0.921809657 0.10234640   1.03767667
#>    65   98 0.507626884 0.03519486 0.999819740 0.01258032   0.56160829
#>    95   32 0.033622129 0.85798781 0.008711212 0.74973864   0.07976456
#>    22   97 0.182670854 0.99544852 0.060067275 0.97954186   0.10076143
#>    26   18 0.271023169 0.72075465 0.266174387 0.81855108   0.22088019
#>    16   30 0.347738310 0.91760896 0.461153054 0.94359077   0.39404484
#>    33   20 0.370835275 0.66654812 0.328980281 0.57967761   0.32173618
#>    15   55 0.054522262 0.46175393 0.016717392 0.38071851   0.09176695
#>    44   94 0.204642149 0.24390424 0.236648898 0.40431577   0.23154347
#>    75   76 0.581373540 0.83562665 0.631598421 0.76201424   0.63054563
#>    11   92 0.489207150 0.45393267 0.573855647 0.61539915   0.47661789
#>    62   51 0.667002992 0.85402435 0.717124235 0.82752652   0.72803251
#>    84   56 0.769235633 0.97972355 0.464599833 1.00000000   0.63720628
#>    49   34 0.514491122 0.35877566 0.556451999 0.40544079   0.50698868
#>    59   83 0.398169243 0.21263363 0.397124396 0.32640190   0.41302300
#>    52    8 0.233194915 0.09662820 0.308203552 0.21655297   0.24175571
#>    50   68 0.846843878 0.88752615 0.875340411 0.78040643   0.85389690
#>    54   89 0.835360153 0.77197338 0.857451691 0.66324642   0.86683121
#>    14   34 0.712180549 0.24891512 0.556451999 0.40544079   0.64183657
#>    89   81 0.857451691 0.66324642 0.868384436 0.60041588   0.88587472
#>    57    3 0.805812576 0.34648104 0.769280914 0.59192468   0.71954561
#>    78   24 0.563535575 0.13422872 0.538764687 0.21147062   0.52380529
#>    80   96 0.920589746 0.65465738 0.911510248 0.51568277   1.14617778
#>    47   84 1.000000000 0.95294441 0.769235633 0.97972355   5.72618897
#>    77   35 0.859573416 0.41478276 0.857967708 0.34844030   0.88596191
#>    23   71 0.949973906 0.70991366 0.999267224 0.29267144   1.99384526
#>    98   47 0.999819740 0.01258032 1.000000000 0.95294441 153.53529033
#>          my1          mx2         my2 bp1 bp2
#>   0.92477390  -9.56650505  3.01023309   0   0
#>   0.95755416   0.01922427  0.92477390   0   0
#>   0.92151011   0.10870757  0.95755416   0   0
#>   0.82940697   0.01922427  0.92477390   0   0
#>   0.79037834   0.01361356  0.60288244   0   0
#>   1.14633234   0.10870757  0.95755416   0   0
#>   0.93413467   0.14716410  0.92151011   0   0
#>   0.80901739   0.11753796  0.80782140   0   0
#>   0.93528910   0.19308904  0.93413467   0   0
#>   0.79718083   0.11753796  0.80782140   0   0
#>   0.80782140   0.15560513  0.64215390   0   0
#>   0.81987643   0.23717256  0.93528910   0   0
#>   0.76728696   0.19367789  0.79718083   0   0
#>   0.70626559   0.30535760  0.66497888   0   0
#>   0.94730742   0.23717256  0.93528910   0   0
#>   0.85750026   0.20888982  0.81987643   0   0
#>   0.85341816   0.38545410  0.87511459   0   0
#>   0.64215390   0.06785549  0.59104897   0   0
#>   0.64229484   0.26063493  0.58960548   0   0
#>   0.77211873   0.32823453  0.70710287   0   0
#>   0.78083472   0.44546293  0.77423015   0   0
#>   0.60288244  -0.26033228  0.55930462   0   0
#>   0.70710287   0.40771629  0.69815117   0   0
#>   0.81052945   0.38545410  0.87511459   0   0
#>   0.69815117   0.44135565  0.60879992   0   0
#>   0.66497888   0.30713111  0.64372305   0   0
#>   0.97600121   0.46959403  0.97138492   0   0
#>   0.91598715   0.38545410  0.87511459   0   0
#>   1.07769969   0.26509629  4.62376808   0   1
#>   0.93796156   0.46959403  0.97138492   0   0
#>   0.89904118   0.40779326  0.91598715   0   0
#>   0.77747774   0.50596724  0.77059920   0   0
#>   0.63668619   0.36024053  0.52999501   0   0
#>   0.39504230   0.08996565  0.37330805   0   0
#>   0.59104897   0.15695793  0.47122756   0   0
#>   0.77423015   0.40771629  0.69815117   0   0
#>   0.47122756   0.09176695  0.39504230   0   0
#>   0.97138492   0.63720628  1.29467935   0   0
#>   0.74098833   0.56134599  0.71592247   0   0
#>   0.60879992   0.44855724  0.60204225   0   0
#>   0.90170813   0.47471830  0.89904118   0   0
#>   0.49900058   0.26063493  0.58960548   0   0
#>   0.97207505   0.50831366  0.93796156   0   0
#>   0.86388619   0.61694004  0.87856131   0   0
#>   0.48561206   0.26951234  0.49900058   0   0
#>   0.77059920   0.52000719  0.79123018   0   0
#>   0.60204225   0.52185505  0.74098833   0   0
#>   0.85322343   0.52000719  0.79123018   0   0
#>   0.47093321   0.23154347  0.32193555   0   0
#>   0.57089855   0.44855724  0.60204225   0   0
#>   0.44844019   0.29747933  0.48048212   0   0
#>   0.56433547   0.46214881  0.57089855   0   0
#>   0.48048212   0.29494070  0.48561206   0   0
#>   0.31513947   0.08996565  0.37330805   0   0
#>   0.81523606   0.67044494  0.79988321   0   0
#>   0.81088840   0.52000719  0.79123018   0   0
#>   0.71592247   0.56753270  0.65869643   0   0
#>   0.65869643   0.47661789  0.56345422   0   0
#>   0.32306988   0.08996565  0.37330805   0   0
#>   0.53950425  -0.26033228  0.55930462   0   0
#>   0.72209962   0.63185935  0.71424030   0   0
#>   1.03256107   0.56418113  0.97207505   0   0
#>   0.87856131   0.65897598  0.96497592   0   0
#>   0.75735092   0.79043401  0.82995495   0   0
#>   0.63591061   0.63185935  0.71424030   0   0
#>   0.31875484   0.31095725  0.30523832   0   0
#>   0.61868595   0.56753270  0.65869643   0   0
#>   0.90881143   0.74768955  0.90208097   0   0
#>   0.37271301   0.29747933  0.48048212   0   0
#>   0.24527556   0.06468681  0.32306988   0   0
#>   0.38737777   0.32045724  0.37271301   0   0
#>   0.39148163   0.43602462  0.44844019   0   0
#>   0.65065782   0.64884882  0.63591061   0   0
#>   0.50642437   0.69590389  0.45062541   0   0
#>   0.50917678   0.50698868  0.40771979   0   0
#>   0.88433106   0.82126298  0.91240902   0   0
#>   0.31262913   0.24175571  0.17469372   0   0
#>   0.15696400   0.11171524  0.22900483   0   0
#>   0.22900483   0.21098513  0.31513947   0   0
#>   0.26965898   0.31095725  0.30523832   0   0
#>   0.74975502   0.75445227  0.67568847   0   0
#>   0.27806336   0.50838354  0.25570883   0   0
#>   0.30094792   0.44345342  0.38737777   0   0
#>   0.17987309   0.31095725  0.30523832   0   0
#>   0.16584820   0.35167382  0.17987309   0   0
#>   0.97048169   0.74768955  0.90208097   0   0
#>   0.95734105   0.91599946  0.99066201   0   0
#>   0.91240902   0.82652709  0.95734105   0   0
#>   0.67503267   0.75445227  0.67568847   0   0
#>   0.45833552   0.69590389  0.45062541   0   0
#>   0.34302666   0.46729374  0.30094792   0   0
#>   0.30869602   0.59896711  0.24841514   0   0
#>   0.82995495   0.81699031  0.71163313   0   0
#>   0.83456971   0.84411606  0.82945023   0   0
#>   0.13831344   0.28744787  0.14611492   0   0
#>   0.25570883   0.52334455  0.16387242   0   0
#>   0.27002113   0.45009615  0.18626184   0   0
#>   0.83205218   0.95904228  0.81623056   0   0
#>   0.72175999   0.88756520  0.71859419   0   0
#>   0.94651221   0.95904228  0.81623056   0   0
#>   0.92348332   0.91051892  0.94651221   0   0
#>   0.25004602   0.50838354  0.25570883   0   0
#>   0.13577555   0.45009615  0.18626184   0   0
#>   0.10313881   0.25917833  0.02955645   0   0
#>   0.06725855   0.25917833  0.02955645   0   0
#>   0.24841514   0.54052475  0.25004602   0   0
#>   0.33465995   0.57893457  0.34302666   0   0
#>   0.14611492   0.26677329  0.10313881   0   0
#>   0.71859419   0.97100574  0.80693600   0   0
#>   0.63582569   0.94014143  0.58359613   0   0
#>   0.62374897   0.77498611  0.67503267   0   0
#>   0.02614828   0.36304579 -0.13776370   0   0
#>  -0.13776370   0.25917833  0.02955645   0   0
#>   0.09350013   0.10596245  0.06567450   0   0
#>   0.12903316   0.08351625  0.24527556   0   0
#>   0.45908162   0.73090103  0.65065782   0   0
#>   0.99066201   5.72618897 42.68770066   0   0
#>   0.70365254   1.14617778  0.57013535   0   0
#>   0.80693600   1.25798239  0.77317650   0   0
#>   0.16408039   0.54052475  0.25004602   0   0
#>   0.23198939   0.70295253  0.19593302   0   0
#>   0.18778904   0.63315186  0.14151873   0   0
#>   0.50259958   0.73088120  0.46076880   0   0
#>   0.52496008   0.81646973  0.62374897   0   0
#>   0.35676215   0.84130245  0.21039621   0   0
#>   0.28081442   0.63581117  0.30869602   0   0
#>   0.08877049   0.65392316 -0.03888767   0   0
#>   0.57013535   1.29981251  0.53971098   0   0
#>   0.21447627   0.83480726  0.26980729   0   0
#>  42.68770066   0.79672806  0.97048169   0   0
#>   0.96960439   1.25798239  0.77317650   0   0
#>   0.48198048   0.91409033  0.45053784   0   0
#>   0.46076880   0.83058160  0.38229380   0   0
#>   0.11217031   0.51937491  0.09386092   0   0
#>  -2.06090739   0.10596245  0.06567450   0   0
#>   0.38095341   0.92456525  0.31028891   0   0
#>   0.26980729   0.83058160  0.38229380   0   0
#>   0.11080711   0.63315186  0.14151873   0   0
#>   0.13923250   0.70295253  0.19593302   0   0
#>   0.09386092   0.63121233 -0.60660656   0   0
#>   0.62170454 153.53529033  0.45352268   0   0
#>   0.17632217   0.65432050  0.28081442   0   0
#>   0.15618082   0.74309860  0.13923250   0   0
#>  -3.07662064   0.87931459 -0.01336328   0   0
#>   0.27157581   0.92456525  0.31028891   0   0
#>   0.21883851   0.90928575  0.27157581   0   0
#>   0.15482694   0.90928575  0.27157581   0   0
#>   0.12425929   0.86462151  0.21883851   0   0
#>  -0.01336328   0.80439113  0.15618082   0   0
#>  -4.15738808   0.57462547 -5.17451976   0   1
#>   0.92151011   0.11313392  0.82940697   0   0
#>   0.82940697   0.11505833  0.80901739   0   0
#>   0.79037834   0.11505833  0.80901739   0   0
#>   3.01023309 -13.71106416  3.93645268   0   1
#>   1.14633234   0.19308904  0.93413467   0   0
#>   0.81987643   0.19367789  0.79718083   0   0
#>   0.76728696   0.21725647  0.70626559   0   0
#>   0.94730742   0.31980527  0.85750026   0   0
#>   0.85750026   0.32476291  0.85341816   0   0
#>   0.64215390   0.15636044  0.64229484   0   0
#>   0.64229484   0.21725647  0.70626559   0   0
#>   0.77211873   0.32463239  0.78083472   0   0
#>   0.78083472   0.32476291  0.85341816   0   0
#>   0.60288244   0.06785549  0.59104897   0   0
#>   0.81052945   0.45078056  0.77747774   0   0
#>   0.66497888   0.32823453  0.70710287   0   0
#>   0.97600121   0.32234422  1.07769969   0   0
#>   1.07769969   0.26085615  0.94730742   0   0
#>   0.93796156   0.49171808  0.90170813   0   0
#>   0.89904118   0.43854596  0.81052945   0   0
#>   0.63668619   0.44135565  0.60879992   0   0
#>   0.64372305   0.26063493  0.58960548   0   0
#>   0.77423015   0.45078056  0.77747774   0   0
#>   0.47122756   0.15877501  0.47093321   0   0
#>   0.74098833   0.50596724  0.77059920   0   0
#>   0.90170813   0.52986927  0.85322343   0   0
#>   0.49900058   0.15877501  0.47093321   0   0
#>   0.97207505   0.54961092  0.86388619   0   0
#>   0.52999501   0.29494070  0.48561206   0   0
#>   0.85322343   0.54961092  0.86388619   0   0
#>   0.57089855   0.36024053  0.52999501   0   0
#>   0.56433547   0.43602462  0.44844019   0   0
#>   0.32193555   0.21098513  0.31513947   0   0
#>   0.81523606   0.61694004  0.87856131   0   0
#>   0.81088840   0.60041615  0.72209962   0   0
#>   0.71592247   0.60041615  0.72209962   0   0
#>   0.32306988  -1.17296918  0.53950425   0   0
#>   0.53950425  -3.32452465  0.57984428   0   1
#>   1.03256107   0.65897598  0.96497592   0   0
#>   0.75735092   0.67044494  0.79988321   0   0
#>   0.63591061   0.63193022  0.61868595   0   0
#>   0.56345422   0.47132070  0.56433547   0   0
#>   0.31875484   0.32045724  0.37271301   0   0
#>   0.71424030   0.70282900  0.74975502   0   0
#>   0.61868595   0.61335695  0.50642437   0   0
#>   0.90881143   0.65897598  0.96497592   0   0
#>   0.39148163   0.44345342  0.38737777   0   0
#>   0.50917678   0.61335695  0.50642437   0   0
#>   1.29467935   0.63330305  1.03256107   0   0
#>   0.88433106   0.74768955  0.90208097   0   0
#>   0.40771979   0.44587587  0.39148163   0   0
#>   0.31262913   0.29425953  0.31875484   0   0
#>   0.15696400   0.12864300  0.09350013   0   0
#>   0.22900483   0.08351625  0.24527556   0   0
#>   0.26965898   0.35167382  0.17987309   0   0
#>   0.74975502   0.70302442  0.75735092   0   0
#>   0.27806336   0.46729374  0.30094792   0   0
#>   0.16584820   0.33131246  0.13831344   0   0
#>   0.97048169   0.82652709  0.95734105   0   0
#>   0.67568847   0.73090103  0.65065782   0   0
#>   0.45833552   0.83480726  0.26980729   0   0
#>   0.82995495   0.79260226  0.83456971   0   0
#>   0.83456971   0.76958599  0.88433106   0   0
#>   0.13831344   0.39047211  0.02614828   0   0
#>   0.17469372   0.15030529  0.15696400   0   0
#>   0.27002113   0.46128712  0.27806336   0   0
#>   0.83205218   0.88047503  0.92348332   0   0
#>   0.82945023   0.86683121  0.72175999   0   0
#>   0.94651221   0.91703888  0.96960439   0   0
#>   0.92348332   0.82126298  0.91240902   0   0
#>   0.13577555   0.48694258  0.11217031   0   0
#>   0.18626184   0.35106283  0.16584820   0   0
#>   0.06725855   0.26677329  0.10313881   0   0
#>   0.71163313   0.77498611  0.67503267   0   0
#>   0.24841514   0.59945496  0.23198939   0   0
#>   0.33465995   0.63581117  0.30869602   0   0
#>   0.16387242   0.48808228  0.13577555   0   0
#>   0.63582569   0.89510161  0.70365254   0   0
#>   0.02614828   0.48694258  0.11217031   0   0
#>  -0.13776370   0.41127232 -2.06090739   0   0
#>   0.09350013   0.20651409  0.06725855   0   0
#>   0.06567450  -0.22085007  0.12903316   0   0
#>   0.12903316  -3.71595240 -0.07719847   0   1
#>   0.45908162   0.71798832  0.45833552   0   0
#>   0.70365254   0.88756520  0.71859419   0   0
#>   0.80693600   0.95904228  0.81623056   0   0
#>   0.45062541   0.66405190  0.35676215   0   0
#>   0.16408039   0.52334455  0.16387242   0   0
#>   0.18778904   0.59945496  0.23198939   0   0
#>   0.50259958   0.91409033  0.45053784   0   0
#>   0.52496008   0.81294765  0.50259958   0   0
#>   0.21039621   0.81747852  0.17632217   0   0
#>   0.28081442   0.70295253  0.19593302   0   0
#>   0.08877049   0.63315186  0.14151873   0   0
#>   0.21447627   0.84130245  0.21039621   0   0
#>   0.96960439   0.91599946  0.99066201   0   0
#>   0.77317650   1.99384526  0.62170454   0   0
#>   0.48198048   0.92456525  0.31028891   0   0
#>   0.58359613   0.82493397  0.52496008   0   0
#>  -2.06090739   0.56160829 -4.15738808   0   0
#>   0.38095341   0.91409033  0.45053784   0   0
#>   0.11080711   0.65392316 -0.03888767   0   0
#>   0.13923250   0.69685834  0.11080711   0   0
#>   0.09386092   0.55132933  0.08877049   0   0
#>   0.17632217   0.80439113  0.15618082   0   0
#>  -0.03888767   0.63121233 -0.60660656   0   0
#>  -0.60660656   0.61126578 -3.07662064   0   0
#>   0.53971098   1.15310560  0.48198048   0   0
#>   0.21883851   0.84780621  0.21447627   0   0
#>   0.45352268   2.11534440  0.15482694   0   0
#>   0.15482694   1.03767667  0.12425929   0   0
#>   0.12425929   0.87931459 -0.01336328   0   0
#>  -4.15738808   0.61126578 -3.07662064   0   0
#>   0.79037834  -9.56650505  3.01023309   0   0
#>   1.14633234  -0.93993335  9.16768067   0   1
#>   0.76728696   0.31833344  0.77211873   0   0
#>   0.97600121   0.40779326  0.91598715   0   0
#>   0.63668619   0.30713111  0.64372305   0   0
#>   0.39504230  -0.26033228  0.55930462   0   0
#>   0.32193555   0.27818539  0.31262913   0   0
#>   0.81523606   0.62417346  0.81088840   0   0
#>   0.56345422   0.58015177  0.50917678   0   0
#>   0.90881143   0.67044494  0.79988321   0   0
#>   1.29467935   0.85473187  4.56280920   0   1
#>   0.40771979   0.57893457  0.34302666   0   0
#>   0.26965898   0.45245562  0.27002113   0   0
#>   0.17469372   0.28744787  0.14611492   0   0
#>   0.83205218   0.84411606  0.82945023   0   0
#>   0.72175999   0.81699031  0.71163313   0   0
#>   0.33465995   0.66405190  0.35676215   0   0
#>   0.63582569   0.81646973  0.62374897   0   0
#>   0.45908162   0.73088120  0.46076880   0   0
#>   0.16408039   0.59773488  0.18778904   0   0
#>   0.57013535   0.94014143  0.58359613   0   0
#>  42.68770066   6.22238083 46.96354457   0   1
#>   0.38095341   0.83058160  0.38229380   0   0
#>   0.62170454   1.29981251  0.53971098   0   0
#>   0.45352268 154.59870818  0.45331883   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.085780755 0.90934941
#>   [2,] 0.765146698 0.03311526
#>   [3,] 0.769280914 0.59192468
#>   [4,] 0.993550233 0.23361283
#>   [5,] 0.973063603 0.90619041
#>   [6,] 0.383225659 0.81834060
#>   [7,] 0.456282136 0.69871753
#>   [8,] 0.308203552 0.21655297
#>   [9,] 0.165582466 0.72701619
#>  [10,] 0.527698348 0.21362291
#>  [11,] 0.489207150 0.45393267
#>  [12,] 0.779055593 0.32990144
#>  [13,] 0.195516273 0.56660089
#>  [14,] 0.712180549 0.24891512
#>  [15,] 0.054522262 0.46175393
#>  [16,] 0.347738310 0.91760896
#>  [17,] 0.825617362 0.97306127
#>  [18,] 0.266174387 0.81855108
#>  [19,] 0.566732253 0.90280052
#>  [20,] 0.328980281 0.57967761
#>  [21,] 0.593333424 0.77225285
#>  [22,] 0.182670854 0.99544852
#>  [23,] 0.949973906 0.70991366
#>  [24,] 0.538764687 0.21147062
#>  [25,] 0.540918712 0.28865955
#>  [26,] 0.271023169 0.72075465
#>  [27,] 0.441586369 0.86631570
#>  [28,] 0.365295705 0.23509557
#>  [29,] 0.016823875 0.00000000
#>  [30,] 0.461153054 0.94359077
#>  [31,] 0.384086693 0.43575161
#>  [32,] 0.008711212 0.74973864
#>  [33,] 0.370835275 0.66654812
#>  [34,] 0.556451999 0.40544079
#>  [35,] 0.857967708 0.34844030
#>  [36,] 0.378788670 0.73716598
#>  [37,] 0.523988381 0.66300066
#>  [38,] 0.597772123 0.08112133
#>  [39,] 0.253545513 0.85578112
#>  [40,] 0.282643632 0.07283984
#>  [41,] 0.475447019 0.85247777
#>  [42,] 0.921809657 0.10234640
#>  [43,] 0.394931773 0.48264437
#>  [44,] 0.204642149 0.24390424
#>  [45,] 0.669941403 0.68539284
#>  [46,] 0.047823732 0.15990143
#>  [47,] 1.000000000 0.95294441
#>  [48,] 0.139567148 0.31890296
#>  [49,] 0.514491122 0.35877566
#>  [50,] 0.846843878 0.88752615
#>  [51,] 0.717124235 0.82752652
#>  [52,] 0.233194915 0.09662820
#>  [53,] 0.543394362 0.90594353
#>  [54,] 0.835360153 0.77197338
#>  [55,] 0.016717392 0.38071851
#>  [56,] 0.464599833 1.00000000
#>  [57,] 0.805812576 0.34648104
#>  [58,] 0.814306253 0.94741108
#>  [59,] 0.398169243 0.21263363
#>  [60,] 0.209977329 0.02773072
#>  [61,] 0.412830924 0.14150496
#>  [62,] 0.667002992 0.85402435
#>  [63,] 0.503425305 0.20966861
#>  [64,] 0.658367104 0.20681622
#>  [65,] 0.507626884 0.03519486
#>  [66,] 0.836121750 0.94485523
#>  [67,] 0.707496939 0.24160196
#>  [68,] 0.875340411 0.78040643
#>  [69,] 0.000000000 0.28512195
#>  [70,] 0.889589371 0.87510046
#>  [71,] 0.999267224 0.29267144
#>  [72,] 0.495857469 0.98379445
#>  [73,] 0.352568135 0.58819035
#>  [74,] 0.774595721 0.75833534
#>  [75,] 0.581373540 0.83562665
#>  [76,] 0.631598421 0.76201424
#>  [77,] 0.859573416 0.41478276
#>  [78,] 0.563535575 0.13422872
#>  [79,] 0.245048732 0.07672028
#>  [80,] 0.920589746 0.65465738
#>  [81,] 0.868384436 0.60041588
#>  [82,] 0.240525224 0.65567552
#>  [83,] 0.397124396 0.32640190
#>  [84,] 0.769235633 0.97972355
#>  [85,] 0.109585001 0.71414906
#>  [86,] 0.185894234 0.87235958
#>  [87,] 0.155328050 0.98355162
#>  [88,] 0.661256009 0.21510864
#>  [89,] 0.857451691 0.66324642
#>  [90,] 0.928446194 0.38694202
#>  [91,] 0.548806580 0.04176966
#>  [92,] 0.573855647 0.61539915
#>  [93,] 0.685851605 0.59686983
#>  [94,] 0.236648898 0.40431577
#>  [95,] 0.033622129 0.85798781
#>  [96,] 0.911510248 0.51568277
#>  [97,] 0.060067275 0.97954186
#>  [98,] 0.999819740 0.01258032
#>  [99,] 0.609150991 0.67220759
#> [100,] 0.163434464 0.36855882
#> 
#> $tri.obj
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,97,95), [6,4,2]
#> 2: (87,97,1), [1,3,7]
#> 3: (86,87,1), [2,4,8]
#> 4: (86,1,95), [1,9,3]
#> 5: (95,32,85), [24,9,6]
#> 6: (32,95,97), [1,0,5]
#> 7: (22,97,87), [2,8,0]
#> 8: (22,87,86), [3,10,7]
#> 9: (95,85,86), [12,4,5]
#> 10: (39,22,86), [8,13,16]
#> 11: (18,86,9), [12,14,13]
#> 12: (86,85,9), [19,11,9]
#> 13: (18,39,86), [10,11,17]
#> 14: (26,18,9), [11,15,21]
#> 15: (9,82,26), [29,14,20]
#> 16: (16,22,39), [10,17,32]
#> 17: (16,39,18), [13,18,16]
#> 18: (18,6,16), [26,17,22]
#> 19: (9,85,13), [39,20,12]
#> 20: (9,13,82), [23,15,19]
#> 21: (18,26,36), [25,22,14]
#> 22: (18,36,6), [40,18,21]
#> 23: (20,82,13), [20,46,37]
#> 24: (85,32,15), [47,39,5]
#> 25: (26,33,36), [28,21,29]
#> 26: (27,16,6), [18,27,31]
#> 27: (41,27,6), [26,35,34]
#> 28: (36,33,7), [44,40,25]
#> 29: (26,82,33), [37,25,15]
#> 30: (16,30,56), [42,32,31]
#> 31: (30,16,27), [26,34,30]
#> 32: (16,56,22), [0,16,30]
#> 33: (53,72,30), [42,45,48]
#> 34: (41,30,27), [31,27,45]
#> 35: (6,7,41), [52,27,40]
#> 36: (33,20,73), [51,44,37]
#> 37: (20,33,82), [29,23,36]
#> 38: (15,55,100), [63,41,47]
#> 39: (85,15,13), [41,19,24]
#> 40: (6,36,7), [28,35,22]
#> 41: (13,15,100), [38,55,39]
#> 42: (30,72,56), [88,30,33]
#> 43: (7,37,21), [66,52,53]
#> 44: (33,73,7), [53,28,36]
#> 45: (53,30,41), [34,54,33]
#> 46: (94,20,13), [23,55,50]
#> 47: (55,15,32), [24,69,38]
#> 48: (19,72,53), [33,49,71]
#> 49: (53,75,19), [72,48,54]
#> 50: (43,20,94), [46,60,51]
#> 51: (43,73,20), [36,50,56]
#> 52: (7,21,41), [57,35,43]
#> 53: (73,37,7), [43,44,56]
#> 54: (53,41,75), [57,49,45]
#> 55: (13,100,94), [62,46,41]
#> 56: (43,37,73), [53,51,59]
#> 57: (21,75,41), [54,52,65]
#> 58: (11,43,31), [60,84,59]
#> 59: (11,37,43), [56,58,75]
#> 60: (31,43,94), [50,81,58]
#> 61: (44,100,48), [63,95,62]
#> 62: (44,94,100), [55,61,93]
#> 63: (48,100,55), [38,68,61]
#> 64: (75,76,62), [80,72,65]
#> 65: (76,75,21), [57,70,64]
#> 66: (21,37,99), [67,70,43]
#> 67: (99,37,92), [75,78,66]
#> 68: (69,48,55), [63,69,82]
#> 69: (69,55,32), [47,0,68]
#> 70: (21,99,76), [77,65,66]
#> 71: (84,72,19), [48,89,88]
#> 72: (75,62,19), [89,49,64]
#> 73: (76,74,51), [111,80,97]
#> 74: (93,45,99), [77,78,85]
#> 75: (11,92,37), [67,59,87]
#> 76: (94,8,83), [92,81,93]
#> 77: (76,99,45), [74,97,70]
#> 78: (93,99,92), [67,86,74]
#> 79: (62,51,84), [98,89,80]
#> 80: (51,62,76), [64,73,79]
#> 81: (83,31,94), [60,76,83]
#> 82: (46,48,69), [68,140,95]
#> 83: (49,31,83), [81,100,84]
#> 84: (49,11,31), [58,83,91]
#> 85: (3,45,93), [74,142,107]
#> 86: (92,34,93), [146,78,87]
#> 87: (92,11,34), [91,86,75]
#> 88: (84,56,72), [42,71,0]
#> 89: (84,19,62), [72,79,71]
#> 90: (51,50,58), [105,98,112]
#> 91: (49,34,11), [87,84,109]
#> 92: (28,83,8), [76,101,96]
#> 93: (94,44,8), [114,76,62]
#> 94: (52,44,46), [95,138,114]
#> 95: (46,44,48), [61,82,94]
#> 96: (59,83,28), [92,101,116]
#> 97: (76,45,74), [107,73,77]
#> 98: (58,84,51), [79,90,103]
#> 99: (83,63,25), [115,100,116]
#> 100: (25,49,83), [83,99,109]
#> 101: (59,28,8), [92,102,96]
#> 102: (61,59,8), [101,113,124]
#> 103: (17,84,58), [98,104,159]
#> 104: (58,66,17), [143,103,105]
#> 105: (50,66,58), [104,90,121]
#> 106: (89,74,3), [107,134,127]
#> 107: (3,74,45), [97,85,106]
#> 108: (57,93,12), [146,170,142]
#> 109: (34,49,25), [100,129,91]
#> 110: (14,25,88), [128,155,129]
#> 111: (51,74,54), [127,112,73]
#> 112: (51,54,50), [118,90,111]
#> 113: (61,8,40), [131,135,102]
#> 114: (52,8,44), [93,94,131]
#> 115: (25,63,10), [130,122,99]
#> 116: (83,59,63), [124,99,96]
#> 117: (50,68,70), [141,121,118]
#> 118: (68,50,54), [112,119,117]
#> 119: (54,89,68), [132,118,127]
#> 120: (66,70,5), [141,160,121]
#> 121: (50,70,66), [120,105,117]
#> 122: (24,25,10), [115,147,128]
#> 123: (78,63,61), [124,166,130]
#> 124: (61,63,59), [116,102,123]
#> 125: (52,79,40), [137,131,126]
#> 126: (52,60,79), [137,125,138]
#> 127: (89,54,74), [111,106,119]
#> 128: (88,25,24), [122,148,110]
#> 129: (14,34,25), [109,110,153]
#> 130: (78,10,63), [115,123,147]
#> 131: (8,52,40), [125,113,114]
#> 132: (89,23,68), [145,119,144]
#> 133: (89,81,80), [164,144,134]
#> 134: (81,89,3), [106,152,133]
#> 135: (61,40,65), [136,166,113]
#> 136: (65,40,60), [137,167,135]
#> 137: (60,40,79), [125,126,136]
#> 138: (52,46,60), [139,126,94]
#> 139: (29,60,46), [138,140,167]
#> 140: (29,46,69), [82,0,139]
#> 141: (68,5,70), [120,117,145]
#> 142: (57,3,93), [85,108,163]
#> 143: (66,47,17), [159,104,160]
#> 144: (89,80,23), [157,132,133]
#> 145: (68,23,5), [161,141,132]
#> 146: (12,93,34), [86,153,108]
#> 147: (78,24,10), [122,130,150]
#> 148: (24,64,88), [149,128,150]
#> 149: (67,88,64), [148,172,155]
#> 150: (24,78,64), [165,148,147]
#> 151: (96,3,77), [163,175,152]
#> 152: (96,81,3), [134,151,164]
#> 153: (34,14,12), [154,146,129]
#> 154: (42,12,14), [153,176,158]
#> 155: (67,14,88), [110,149,176]
#> 156: (78,91,38), [178,165,173]
#> 157: (80,96,23), [181,144,164]
#> 158: (42,35,12), [170,154,184]
#> 159: (47,84,17), [103,143,0]
#> 160: (66,5,47), [161,143,120]
#> 161: (47,5,23), [145,174,160]
#> 162: (71,96,90), [175,182,181]
#> 163: (3,57,77), [169,151,142]
#> 164: (96,80,81), [133,152,157]
#> 165: (38,64,78), [150,156,171]
#> 166: (61,65,78), [173,123,135]
#> 167: (65,60,29), [139,189,136]
#> 168: (77,35,90), [182,175,169]
#> 169: (35,77,57), [163,170,168]
#> 170: (12,35,57), [169,108,158]
#> 171: (2,64,38), [165,178,172]
#> 172: (2,67,64), [149,171,177]
#> 173: (78,65,91), [179,156,166]
#> 174: (23,71,47), [185,161,181]
#> 175: (77,90,96), [162,151,168]
#> 176: (42,14,67), [155,177,154]
#> 177: (42,67,2), [172,188,176]
#> 178: (2,38,91), [156,179,171]
#> 179: (2,91,65), [173,180,178]
#> 180: (65,98,2), [188,179,189]
#> 181: (71,23,96), [157,162,174]
#> 182: (71,90,35), [168,183,162]
#> 183: (4,71,35), [182,184,186]
#> 184: (42,4,35), [183,158,187]
#> 185: (98,47,71), [174,186,0]
#> 186: (98,71,4), [183,187,185]
#> 187: (98,4,42), [184,188,186]
#> 188: (98,42,2), [177,180,187]
#> 189: (65,29,98), [0,180,167]
#> boundary nodes:  97 32 69 29 98 47 84 56 22 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.124397480 0.074754015 0.101699526 0.073202026 0.106967938 0.095345128 
#>                                                                         
#> 0.115316787 0.175652121 0.123131145 0.146755812 0.057456759 0.069653001 
#>                                                                         
#> 0.136005039 0.071870844 0.156621197 0.039313673 0.105420871 0.170747314 
#>                                                                         
#> 0.099800273 0.109007958 0.081295792 0.291605640 0.071064327 0.075548547 
#>                                                                         
#> 0.091301902 0.130762850 0.056514437 0.106950721 0.281965716 0.053110570 
#>                                                                         
#> 0.079713811 0.154950003 0.025076956 0.147220098 0.175704859 0.086507341 
#>                                                                         
#> 0.143342931 0.035208799 0.129401607 0.151567975 0.092227414 0.134103102 
#>                                                                         
#> 0.091216902 0.068750974 0.198183903 0.142594678 0.076549530 0.107258511 
#>                                                                         
#> 0.081479523 0.187033373 0.048130522 0.221774658 0.176603621 0.055094037 
#>                                                                         
#> 0.098586725 0.064492458 0.085658844 0.068939451 0.137525358 0.369106976 
#>                                                                         
#> 0.092569553 0.107625518 0.111504871 0.089946608 0.062203900 0.141328352 
#>                                                                         
#> 0.066880192 0.160871164 0.150751835 0.143597165 0.110124207 0.106681121 
#>                                                                         
#> 0.089941412 0.231061329 0.082905564 0.068153510 0.107112352 0.177900219 
#>                                                                         
#> 0.099295338 0.096694942 0.127566192 0.087437544 0.121749780 0.060027841 
#>                                                                         
#> 0.090051024 0.055456799 0.030098553 0.021964705 0.166495508 0.282770411 
#>                                                                         
#> 0.074931219 0.141034750 0.062276100 0.116122002 0.145968400 0.024593031 
#>                                                                         
#> 0.105297816 0.095760072 0.118517822 0.089075993 0.087889220 0.076192352 
#>                                                                         
#> 0.113373824 0.037794633 0.060249311 0.077218975 0.117809769 0.054873144 
#>                                                                         
#> 0.102661531 0.075283034 0.113405808 0.228110955 0.085529082 0.209195803 
#>                                                                         
#> 0.134042112 0.083575739 0.175539148 0.062583425 0.197630204 0.011273695 
#>                                                                         
#> 0.008781232 0.119423465 0.198826541 0.099466620 0.105028831 0.061124359 
#>                                                                         
#> 0.062818617 0.198002757 0.081060540 0.056773983 0.053958375 0.129849938 
#>                                                                         
#> 0.086921533 0.113725487 0.195133927 0.080309398 0.052191919 0.139538367 
#>                                                                         
#> 0.060197914 0.041701264 0.660273376 0.008684384 0.216310421 0.235569779 
#>                                                                         
#> 0.151906940 0.177673769 0.059334673 0.149591396 0.171278233 0.983076363 
#>                                                                         
#> 0.106728424 0.152948821 0.162665127 0.235471787 0.029818875 0.096645007 
#>                                                                         
#> 0.105626458 0.112671931 0.128316552 0.163184253 0.103469360 0.138944275 
#>                                                                         
#> 0.117051461 0.258331585 0.098336799 0.113581710 0.142985660 0.182500034 
#>                                                                         
#> 0.090448592 0.036579119 0.080458854 0.116618948 0.140167493 0.200623790 
#>                                                                         
#> 0.155532950 0.086460559 0.167416708 0.023548580 0.117324541 0.079917964 
#>                                                                         
#> 0.113730524 0.098550538 0.131289137 0.087583550 0.039611097 0.101287953 
#>                                                                         
#> 0.097047272 0.464698351 0.216621261 0.107733578 0.107511683 0.211941402 
#>                                                                         
#> 0.178389901 0.085679773 0.113518414 0.162024079 0.151428567 0.210678431 
#>                                                                         
#> 0.273408473 0.154326440 0.098458800 0.200935154 0.195872347 0.183571081 
#>                                                                         
#> 0.039814670 0.143044616 0.148665027 0.128759817 0.028033441 0.136398814 
#>                                                                         
#> 0.031477282 0.130636458 0.142923550 0.147185701 0.150018303 0.157881426 
#>                                                                         
#> 0.044514886 0.040859972 0.142295607 0.058323129 0.150880203 0.072624049 
#>                                                                         
#> 0.023169777 0.126123007 0.122545335 0.175813071 0.096459429 0.063719586 
#>                                                                         
#> 0.142436344 0.297743129 0.072704324 0.162878661 0.285617871 0.277642183 
#>                                                                         
#> 0.103625244 0.159284121 0.235071377 0.087107663 0.119692947 0.113482356 
#>                                                                         
#> 0.161375377 0.255786537 0.053292775 0.063186514 0.268626170 0.164077773 
#>                                                                         
#> 0.248126093 0.117909117 0.095076471 0.492063280 0.074287056 0.174123034 
#>                                                                         
#> 0.203896794 0.093624895 0.255581791 0.216513152 0.257528211 0.239656708 
#>                                                                         
#> 0.254240044 0.280091671 0.221121414 0.118926547 0.492712113 0.111078510 
#>                                                                         
#> 0.123631144 0.097916557 0.116352734 0.096427827 0.089420062 0.163573502 
#>                                                                         
#> 0.089114110 0.182309605 0.056694564 0.305309852 0.062756267 0.113773062 
#>                                                                         
#> 0.141450507 0.110845330 0.110948579 0.220797791 0.063774623 0.248147416 
#>                                                                         
#> 0.081116635 0.139270889 0.232312968 0.066361893 0.420143900 0.940364113 
#> 
#> $alpha
#> [1] 0.189391
#> 
#> attr(,"class")
#> [1] "scree"