Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1
#>    67    6 0.32326468 0.29756133 0.26922383 0.32346001  0.299118411
#>    64   67 0.38008978 0.35491849 0.32326468 0.29756133  0.355512380
#>     1   86 0.29284031 0.28646461 0.25841619 0.20290179  0.331372757
#>    67    1 0.32326468 0.29756133 0.29284031 0.28646461  0.332955917
#>     1   40 0.29284031 0.28646461 0.17220296 0.27438668  0.234203552
#>     6   40 0.26922383 0.32346001 0.17220296 0.27438668  0.233126774
#>     6  100 0.26922383 0.32346001 0.37112278 0.45043173  0.274999567
#>    64  100 0.38008978 0.35491849 0.37112278 0.45043173  0.308432023
#>    94    9 0.46442086 0.43282396 0.43566510 0.51345839  0.421819683
#>    17   67 0.39370191 0.30828270 0.32326468 0.29756133  0.366375166
#>    45   40 0.20671113 0.20702213 0.17220296 0.27438668  0.188012173
#>    40   58 0.17220296 0.27438668 0.04374852 0.24057473  0.098148791
#>    20   86 0.33907371 0.09516107 0.25841619 0.20290179  0.223625143
#>    69   20 0.16111061 0.19671328 0.33907371 0.09516107  0.211746421
#>    45   97 0.20671113 0.20702213 0.15017002 0.23828897  0.178987717
#>    78   68 0.17237344 0.35920075 0.12448565 0.37942849  0.126283841
#>   100   90 0.37112278 0.45043173 0.38299380 0.49859433  0.303980405
#>    78    6 0.17237344 0.35920075 0.26922383 0.32346001  0.248825633
#>    17   94 0.39370191 0.30828270 0.46442086 0.43282396  0.460054594
#>    26   82 0.40339940 0.19960766 0.54558254 0.32276330  0.474863927
#>    20   26 0.33907371 0.09516107 0.40339940 0.19960766  0.330257246
#>    37   90 0.22685618 0.51074065 0.38299380 0.49859433  0.308155591
#>     9   90 0.43566510 0.51345839 0.38299380 0.49859433  0.421299962
#>    94   82 0.46442086 0.43282396 0.54558254 0.32276330  0.541553901
#>    26   60 0.40339940 0.19960766 0.47995877 0.09142809  0.510579879
#>    68   37 0.12448565 0.37942849 0.22685618 0.51074065  0.168026766
#>    98   37 0.16916287 0.53484910 0.22685618 0.51074065  0.211615716
#>    51   18 0.07570866 0.19482753 0.05755845 0.14590559  0.119402860
#>    69   97 0.16111061 0.19671328 0.15017002 0.23828897  0.118147656
#>    61   51 0.05837684 0.22413707 0.07570866 0.19482753  0.104233141
#>    58   61 0.04374852 0.24057473 0.05837684 0.22413707  0.097698701
#>    68   58 0.12448565 0.37942849 0.04374852 0.24057473  0.095964356
#>    49   38 0.50604875 0.48090678 0.45011069 0.51006492  0.466433275
#>    58   18 0.04374852 0.24057473 0.05755845 0.14590559 -0.001428372
#>    98   76 0.16916287 0.53484910 0.08946118 0.53705827  0.127283132
#>     9   46 0.43566510 0.51345839 0.48244126 0.52532576  0.438264256
#>     9   38 0.43566510 0.51345839 0.45011069 0.51006492  0.451590397
#>    46   49 0.48244126 0.52532576 0.50604875 0.48090678  0.504870469
#>    99   71 0.55469345 0.48577354 0.52288707 0.48748473  0.535731033
#>    88   32 0.38787906 0.67337537 0.23497463 0.59630813  0.335346152
#>    20   60 0.33907371 0.09516107 0.47995877 0.09142809  0.410311789
#>    27   20 0.06716017 0.08457707 0.33907371 0.09516107  0.204073553
#>    12   58 0.00000000 0.42815704 0.04374852 0.24057473  0.036391768
#>    76    8 0.08946118 0.53705827 0.04065085 0.47077173  0.122419267
#>    32   66 0.23497463 0.59630813 0.07210092 0.65792558  0.150961090
#>    46   71 0.48244126 0.52532576 0.52288707 0.48748473  0.542123241
#>    49   94 0.50604875 0.48090678 0.46442086 0.43282396  0.537551287
#>     8   12 0.04065085 0.47077173 0.00000000 0.42815704  0.063863397
#>    66   12 0.07210092 0.65792558 0.00000000 0.42815704 -0.133570244
#>    18   27 0.05755845 0.14590559 0.06716017 0.08457707  0.131530714
#>    18   12 0.05755845 0.14590559 0.00000000 0.42815704 -3.709231785
#>    60   57 0.47995877 0.09142809 0.54607596 0.08708244  0.518811513
#>    20   30 0.33907371 0.09516107 0.49592643 0.02967131  0.408101629
#>    30   57 0.49592643 0.02967131 0.54607596 0.08708244  0.511531523
#>    43   99 0.61743321 0.57227221 0.55469345 0.48577354  0.599409132
#>    57   93 0.54607596 0.08708244 0.50179539 0.00000000  0.559496617
#>    88   11 0.38787906 0.67337537 0.30280813 0.70685540  0.318257145
#>     9   88 0.43566510 0.51345839 0.38787906 0.67337537  0.386775403
#>    79   88 0.57172184 0.65615089 0.38787906 0.67337537  0.476123188
#>    30   93 0.49592643 0.02967131 0.50179539 0.00000000  0.557556154
#>    24   57 0.61049590 0.15512799 0.54607596 0.08708244  0.572339690
#>    66   76 0.07210092 0.65792558 0.08946118 0.53705827  0.131218537
#>    43   79 0.61743321 0.57227221 0.57172184 0.65615089  0.537925552
#>    24   70 0.61049590 0.15512799 0.72742187 0.31842528  0.635081317
#>    11   89 0.30280813 0.70685540 0.25835098 0.74429276  0.235516366
#>    75   16 0.75993996 0.39696889 0.65469759 0.52768398  0.644500142
#>    82   70 0.54558254 0.32276330 0.72742187 0.31842528  0.638380705
#>    77   25 0.41546951 0.72242299 0.37125115 0.75775995  0.372677627
#>    25   73 0.37125115 0.75775995 0.32776754 0.70887873  0.371296179
#>    85   70 0.73946283 0.30204453 0.72742187 0.31842528  0.659459489
#>    95   66 0.11014682 0.73036227 0.07210092 0.65792558  0.164345748
#>    77   88 0.41546951 0.72242299 0.38787906 0.67337537  0.478838195
#>     3    4 0.60853147 0.09587632 0.62197573 0.01685974  0.663035115
#>     3   57 0.60853147 0.09587632 0.54607596 0.08708244  0.583019697
#>    85   74 0.73946283 0.30204453 0.82272604 0.33060215  0.773167422
#>    89   73 0.25835098 0.74429276 0.32776754 0.70887873  0.313043785
#>    36   53 0.54313422 0.74697582 0.50525146 0.82456352  0.474176570
#>    89   95 0.25835098 0.74429276 0.11014682 0.73036227  0.189703400
#>    95   52 0.11014682 0.73036227 0.24688819 0.83428633  0.169632427
#>     7   24 0.75170758 0.25090001 0.61049590 0.15512799  0.704496098
#>    15   16 0.69494257 0.60295763 0.65469759 0.52768398  0.743301226
#>    70   75 0.72742187 0.31842528 0.75993996 0.39696889  0.639733504
#>    85   39 0.73946283 0.30204453 0.78337233 0.24628383  0.788295016
#>    10   28 0.09518253 0.74628051 0.08429426 0.92694368 -0.333587563
#>    25   52 0.37125115 0.75775995 0.24688819 0.83428633  0.309374563
#>    15   35 0.69494257 0.60295763 0.66169000 0.73542987  0.647056908
#>    43   16 0.61743321 0.57227221 0.65469759 0.52768398  0.662381090
#>     7   39 0.75170758 0.25090001 0.78337233 0.24628383  0.772545412
#>    62   39 0.78536762 0.21827123 0.78337233 0.24628383  0.764960039
#>    24   80 0.61049590 0.15512799 0.72582444 0.06806990  0.706642018
#>    77   79 0.41546951 0.72242299 0.57172184 0.65615089  0.489770312
#>    24    3 0.61049590 0.15512799 0.60853147 0.09587632  0.676967130
#>    19   54 0.49609981 0.87214054 0.29785683 0.87839438  0.396360418
#>    74   41 0.82272604 0.33060215 0.82174595 0.48749968  0.839292481
#>    74   75 0.82272604 0.33060215 0.75993996 0.39696889  0.772350740
#>    74   39 0.82272604 0.33060215 0.78337233 0.24628383  0.852172084
#>    95   72 0.11014682 0.73036227 0.23835248 0.84756675  0.169358728
#>    53   25 0.50525146 0.82456352 0.37125115 0.75775995  0.436684524
#>    36   35 0.54313422 0.74697582 0.66169000 0.73542987  0.604843982
#>    96   62 0.88027893 0.02115191 0.78536762 0.21827123  0.835753934
#>    62   80 0.78536762 0.21827123 0.72582444 0.06806990  0.804690434
#>    91   80 0.76999873 0.03364192 0.72582444 0.06806990  0.716539282
#>    80    4 0.72582444 0.06806990 0.62197573 0.01685974  0.698422347
#>    65   35 0.69719660 0.81842084 0.66169000 0.73542987  0.778991473
#>    10   95 0.09518253 0.74628051 0.11014682 0.73036227  0.165579606
#>    72   54 0.23835248 0.84756675 0.29785683 0.87839438  0.268723130
#>    93   23 0.50179539 0.00000000 0.77413074 0.01908522  0.714058466
#>    72   52 0.23835248 0.84756675 0.24688819 0.83428633  0.270310102
#>    53   21 0.50525146 0.82456352 0.57004062 0.82062904  0.539665571
#>    63   72 0.22163200 0.86405291 0.23835248 0.84756675  0.246524797
#>    53   19 0.50525146 0.82456352 0.49609981 0.87214054  0.417697278
#>    59   48 0.92347183 0.26108568 0.92679433 0.33089188  0.874850291
#>    23   96 0.77413074 0.01908522 0.88027893 0.02115191  0.826780850
#>    41   48 0.82174595 0.48749968 0.92679433 0.33089188  0.916680462
#>    21   65 0.57004062 0.82062904 0.69719660 0.81842084  0.633827827
#>    21   35 0.57004062 0.82062904 0.66169000 0.73542987  0.633222236
#>    41   13 0.82174595 0.48749968 0.89978632 0.54353396  0.818160761
#>    28   12 0.08429426 0.92694368 0.00000000 0.42815704 -0.899998732
#>    10   63 0.09518253 0.74628051 0.22163200 0.86405291  0.127026602
#>    63   42 0.22163200 0.86405291 0.26114436 0.89445399  0.181291749
#>     5   28 0.33618124 0.96480794 0.08429426 0.92694368  0.196261968
#>    42   54 0.26114436 0.89445399 0.29785683 0.87839438  0.298484352
#>    55    5 0.56206674 0.92642143 0.33618124 0.96480794  0.456284752
#>    96   59 0.88027893 0.02115191 0.92347183 0.26108568  0.883975792
#>    41   22 0.82174595 0.48749968 0.86139441 0.65084490  0.791263824
#>    44   55 0.65497046 0.89274065 0.56206674 0.92642143  0.596272887
#>    55   19 0.56206674 0.92642143 0.49609981 0.87214054  0.551197765
#>     2   44 0.73503347 0.90076192 0.65497046 0.89274065  0.697885676
#>    29   59 1.00000000 0.18857535 0.92347183 0.26108568  0.885254229
#>    14   29 0.93133180 0.02348915 1.00000000 0.18857535  0.900736547
#>    96   93 0.88027893 0.02115191 0.50179539 0.00000000  1.881777704
#>    29   47 1.00000000 0.18857535 0.97184938 0.67445584  1.336882079
#>    29   48 1.00000000 0.18857535 0.92679433 0.33089188  1.024669466
#>    87   55 0.73579548 0.91535529 0.56206674 0.92642143  0.656671375
#>     2   87 0.73503347 0.90076192 0.73579548 0.91535529  0.693650640
#>    65   22 0.69719660 0.81842084 0.86139441 0.65084490  0.803863818
#>    87   81 0.73579548 0.91535529 0.87521737 0.94136436  0.800133900
#>    22   47 0.86139441 0.65084490 0.97184938 0.67445584  0.907659785
#>     2   65 0.73503347 0.90076192 0.69719660 0.81842084  0.806931607
#>    22   13 0.86139441 0.65084490 0.89978632 0.54353396  0.927061002
#>    47   13 0.97184938 0.67445584 0.89978632 0.54353396  1.185169592
#>    22   31 0.86139441 0.65084490 0.94521163 0.76472111  0.861608856
#>    31   34 0.94521163 0.76472111 0.99726093 0.72109866  0.966964551
#>    31   47 0.94521163 0.76472111 0.97184938 0.67445584  0.949443061
#>    84   83 0.97152019 0.78990755 0.96981713 0.94365189  0.964966598
#>    84   33 0.97152019 0.78990755 0.91613595 0.80706641  0.942426689
#>    84   31 0.97152019 0.78990755 0.94521163 0.76472111  0.970048216
#>     2   33 0.73503347 0.90076192 0.91613595 0.80706641  0.824936907
#>    81   83 0.87521737 0.94136436 0.96981713 0.94365189  0.922117186
#>    92   83 0.99417186 0.75919053 0.96981713 0.94365189  1.110009439
#>    56   81 0.86378672 1.00000000 0.87521737 0.94136436  0.912667898
#>     1    6 0.29284031 0.28646461 0.26922383 0.32346001  0.299118411
#>    26   86 0.40339940 0.19960766 0.25841619 0.20290179  0.331372757
#>    26    1 0.40339940 0.19960766 0.29284031 0.28646461  0.332955917
#>    37  100 0.22685618 0.51074065 0.37112278 0.45043173  0.274999567
#>    45    1 0.20671113 0.20702213 0.29284031 0.28646461  0.236997212
#>   100    9 0.37112278 0.45043173 0.43566510 0.51345839  0.421819683
#>    26   67 0.40339940 0.19960766 0.32326468 0.29756133  0.366375166
#>    97   40 0.15017002 0.23828897 0.17220296 0.27438668  0.188012173
#>    97   58 0.15017002 0.23828897 0.04374852 0.24057473  0.098148791
#>    45   86 0.20671113 0.20702213 0.25841619 0.20290179  0.223625143
#>    45   20 0.20671113 0.20702213 0.33907371 0.09516107  0.211746421
#>    40   68 0.17220296 0.27438668 0.12448565 0.37942849  0.126283841
#>    37    6 0.22685618 0.51074065 0.26922383 0.32346001  0.248825633
#>    64   94 0.38008978 0.35491849 0.46442086 0.43282396  0.460054594
#>    17   82 0.39370191 0.30828270 0.54558254 0.32276330  0.474863927
#>    37   78 0.22685618 0.51074065 0.17237344 0.35920075  0.179254565
#>    32   90 0.23497463 0.59630813 0.38299380 0.49859433  0.308155591
#>    38   94 0.45011069 0.51006492 0.46442086 0.43282396  0.432331665
#>    99   82 0.55469345 0.48577354 0.54558254 0.32276330  0.541553901
#>    24   60 0.61049590 0.15512799 0.47995877 0.09142809  0.510579879
#>    32   37 0.23497463 0.59630813 0.22685618 0.51074065  0.211615716
#>    69   18 0.16111061 0.19671328 0.05755845 0.14590559  0.119402860
#>    51   97 0.07570866 0.19482753 0.15017002 0.23828897  0.118147656
#>    97   61 0.15017002 0.23828897 0.05837684 0.22413707  0.097698701
#>    61   18 0.05837684 0.22413707 0.05755845 0.14590559 -0.001428372
#>    68   76 0.12448565 0.37942849 0.08946118 0.53705827  0.127283132
#>    88   46 0.38787906 0.67337537 0.48244126 0.52532576  0.438264256
#>    46   38 0.48244126 0.52532576 0.45011069 0.51006492  0.451590397
#>    71   49 0.52288707 0.48748473 0.50604875 0.48090678  0.504870469
#>    68    8 0.12448565 0.37942849 0.04065085 0.47077173  0.122419267
#>    98   66 0.16916287 0.53484910 0.07210092 0.65792558  0.150961090
#>    99   94 0.55469345 0.48577354 0.46442086 0.43282396  0.537551287
#>    68   12 0.12448565 0.37942849 0.00000000 0.42815704  0.063863397
#>    93   27 0.50179539 0.00000000 0.06716017 0.08457707  0.218222004
#>    69   27 0.16111061 0.19671328 0.06716017 0.08457707  0.131530714
#>    27   12 0.06716017 0.08457707 0.00000000 0.42815704 -3.709231785
#>    24   26 0.61049590 0.15512799 0.40339940 0.19960766  0.514924003
#>    60   30 0.47995877 0.09142809 0.49592643 0.02967131  0.408101629
#>    16   99 0.65469759 0.52768398 0.55469345 0.48577354  0.599409132
#>     4   93 0.62197573 0.01685974 0.50179539 0.00000000  0.559496617
#>    32   11 0.23497463 0.59630813 0.30280813 0.70685540  0.318257145
#>    90   88 0.38299380 0.49859433 0.38787906 0.67337537  0.386775403
#>    93   20 0.50179539 0.00000000 0.33907371 0.09516107  0.388521581
#>    73   88 0.32776754 0.70887873 0.38787906 0.67337537  0.321614599
#>    46   79 0.48244126 0.52532576 0.57172184 0.65615089  0.537925552
#>    46   43 0.48244126 0.52532576 0.61743321 0.57227221  0.546948051
#>     8   66 0.04065085 0.47077173 0.07210092 0.65792558 -0.038716678
#>    32   89 0.23497463 0.59630813 0.25835098 0.74429276  0.235516366
#>    99   70 0.55469345 0.48577354 0.72742187 0.31842528  0.638380705
#>    88   25 0.38787906 0.67337537 0.37125115 0.75775995  0.372677627
#>     7   70 0.75170758 0.25090001 0.72742187 0.31842528  0.659459489
#>    73   11 0.32776754 0.70887873 0.30280813 0.70685540  0.310930674
#>    15   43 0.69494257 0.60295763 0.61743321 0.57227221  0.636593520
#>     4   57 0.62197573 0.01685974 0.54607596 0.08708244  0.583019697
#>    70   74 0.72742187 0.31842528 0.82272604 0.33060215  0.773167422
#>    77   53 0.41546951 0.72242299 0.50525146 0.82456352  0.474176570
#>    32   95 0.23497463 0.59630813 0.11014682 0.73036227  0.189703400
#>    62   24 0.78536762 0.21827123 0.61049590 0.15512799  0.704496098
#>    41   16 0.82174595 0.48749968 0.65469759 0.52768398  0.743301226
#>    99   75 0.55469345 0.48577354 0.75993996 0.39696889  0.639733504
#>    66   28 0.07210092 0.65792558 0.08429426 0.92694368 -0.333587563
#>    89   52 0.25835098 0.74429276 0.24688819 0.83428633  0.309374563
#>    79   35 0.57172184 0.65615089 0.66169000 0.73542987  0.647056908
#>    36   79 0.54313422 0.74697582 0.57172184 0.65615089  0.489770312
#>    80    3 0.72582444 0.06806990 0.60853147 0.09587632  0.676967130
#>    25   54 0.37125115 0.75775995 0.29785683 0.87839438  0.396360418
#>    75   41 0.75993996 0.39696889 0.82174595 0.48749968  0.839292481
#>    59   39 0.92347183 0.26108568 0.78337233 0.24628383  0.852172084
#>    91   62 0.76999873 0.03364192 0.78536762 0.21827123  0.835753934
#>    23   80 0.77413074 0.01908522 0.72582444 0.06806990  0.716539282
#>    23    4 0.77413074 0.01908522 0.62197573 0.01685974  0.698422347
#>    22   35 0.86139441 0.65084490 0.66169000 0.73542987  0.778991473
#>    63   95 0.22163200 0.86405291 0.11014682 0.73036227  0.165579606
#>    59   62 0.92347183 0.26108568 0.78536762 0.21827123  0.855151144
#>    54   52 0.29785683 0.87839438 0.24688819 0.83428633  0.270310102
#>    19   21 0.49609981 0.87214054 0.57004062 0.82062904  0.539665571
#>    21   36 0.57004062 0.82062904 0.54313422 0.74697582  0.535752406
#>    42   72 0.26114436 0.89445399 0.23835248 0.84756675  0.246524797
#>     5   19 0.33618124 0.96480794 0.49609981 0.87214054  0.397316551
#>    25   19 0.37125115 0.75775995 0.49609981 0.87214054  0.417697278
#>    74   48 0.82272604 0.33060215 0.92679433 0.33089188  0.874850291
#>    91   96 0.76999873 0.03364192 0.88027893 0.02115191  0.826780850
#>    13   48 0.89978632 0.54353396 0.92679433 0.33089188  0.916680462
#>    44   65 0.65497046 0.89274065 0.69719660 0.81842084  0.633827827
#>    22   15 0.86139441 0.65084490 0.69494257 0.60295763  0.760095433
#>    28   63 0.08429426 0.92694368 0.22163200 0.86405291  0.127026602
#>    28   42 0.08429426 0.92694368 0.26114436 0.89445399  0.181291749
#>     5   54 0.33618124 0.96480794 0.29785683 0.87839438  0.298484352
#>    21   55 0.57004062 0.82062904 0.56206674 0.92642143  0.596272887
#>    96   29 0.88027893 0.02115191 1.00000000 0.18857535  0.900736547
#>    14   93 0.93133180 0.02348915 0.50179539 0.00000000  1.881777704
#>    48   47 0.92679433 0.33089188 0.97184938 0.67445584  1.336882079
#>    44   87 0.65497046 0.89274065 0.73579548 0.91535529  0.693650640
#>    33   22 0.91613595 0.80706641 0.86139441 0.65084490  0.803863818
#>    33   65 0.91613595 0.80706641 0.69719660 0.81842084  0.806931607
#>    33   31 0.91613595 0.80706641 0.94521163 0.76472111  0.861608856
#>    81    2 0.87521737 0.94136436 0.73503347 0.90076192  0.810020326
#>    92   34 0.99417186 0.75919053 0.99726093 0.72109866  0.966964551
#>    34   47 0.99726093 0.72109866 0.97184938 0.67445584  0.949443061
#>    34   29 0.99726093 0.72109866 1.00000000 0.18857535  1.426432396
#>    33   83 0.91613595 0.80706641 0.96981713 0.94365189  0.964966598
#>    92   31 0.99417186 0.75919053 0.94521163 0.76472111  0.970048216
#>    56   87 0.86378672 1.00000000 0.73579548 0.91535529  0.664322501
#>    81   33 0.87521737 0.94136436 0.91613595 0.80706641  0.824936907
#>    50   83 0.96560657 0.98397542 0.96981713 0.94365189  0.922117186
#>    56   55 0.86378672 1.00000000 0.56206674 0.92642143  0.559223335
#>    50   81 0.96560657 0.98397542 0.87521737 0.94136436  0.912667898
#>    50   92 0.96560657 0.98397542 0.99417186 0.75919053  5.106416930
#>    64    6 0.38008978 0.35491849 0.26922383 0.32346001  0.319201681
#>    17   64 0.39370191 0.30828270 0.38008978 0.35491849  0.355512380
#>    78   40 0.17237344 0.35920075 0.17220296 0.27438668  0.211714617
#>   100   94 0.37112278 0.45043173 0.46442086 0.43282396  0.421819683
#>    26   17 0.40339940 0.19960766 0.39370191 0.30828270  0.366375166
#>    45   69 0.20671113 0.20702213 0.16111061 0.19671328  0.211746421
#>    98   68 0.16916287 0.53484910 0.12448565 0.37942849  0.168026766
#>    32   98 0.23497463 0.59630813 0.16916287 0.53484910  0.211615716
#>    69   51 0.16111061 0.19671328 0.07570866 0.19482753  0.119402860
#>    49   99 0.50604875 0.48090678 0.55469345 0.48577354  0.535731033
#>    99   46 0.55469345 0.48577354 0.48244126 0.52532576  0.542123241
#>    24   82 0.61049590 0.15512799 0.54558254 0.32276330  0.514924003
#>     7   85 0.75170758 0.25090001 0.73946283 0.30204453  0.659459489
#>    15   79 0.69494257 0.60295763 0.57172184 0.65615089  0.636593520
#>    25   89 0.37125115 0.75775995 0.25835098 0.74429276  0.313043785
#>    77   36 0.41546951 0.72242299 0.54313422 0.74697582  0.474176570
#>    62    7 0.78536762 0.21827123 0.75170758 0.25090001  0.704496098
#>    41   15 0.82174595 0.48749968 0.69494257 0.60295763  0.743301226
#>    66   10 0.07210092 0.65792558 0.09518253 0.74628051 -0.333587563
#>    59   74 0.92347183 0.26108568 0.82272604 0.33060215  0.852172084
#>    23   91 0.77413074 0.01908522 0.76999873 0.03364192  0.716539282
#>    44   21 0.65497046 0.89274065 0.57004062 0.82062904  0.633827827
#>    42    5 0.26114436 0.89445399 0.33618124 0.96480794  0.196261968
#>    96   14 0.88027893 0.02115191 0.93133180 0.02348915  0.900736547
#>    92   84 0.99417186 0.75919053 0.97152019 0.78990755  0.970048216
#>    56    5 0.86378672 1.00000000 0.33618124 0.96480794  0.559223335
#>    50   56 0.96560657 0.98397542 0.86378672 1.00000000  0.912667898
#>            my1          mx2           my2 bp1 bp2
#>    0.316507950  0.319201681   0.358414212   0   0
#>    0.322440340  0.319201681   0.358414212   0   0
#>    0.221718972  0.236997212   0.260597448   0   0
#>    0.223734155  0.299118411   0.316507950   0   0
#>    0.263626251  0.233126774   0.274381371   0   0
#>    0.274381371  0.211714617   0.316714470   0   0
#>    0.423199271  0.308432023   0.396368635   0   0
#>    0.396368635  0.411048393   0.406002499   0   0
#>    0.463076212  0.432331665   0.466824983   0   0
#>    0.251074040  0.355512380   0.322440340   0   0
#>    0.239964253  0.234203552   0.263626251   0   0
#>    0.294814105  0.095964356   0.303112958   0   0
#>    0.092794776  0.330257246   0.172622359   0   0
#>    0.078738962  0.204073553   0.065292800   0   0
#>    0.223644958  0.188012173   0.239964253   0   0
#>    0.316886184  0.179254565   0.442290763   0   0
#>    0.492525123  0.421299962   0.463608431   0   0
#>    0.417278059  0.211714617   0.316714470   0   0
#>    0.352954288  0.466417957   0.349340944   0   0
#>    0.260754906  0.514924003   0.214505564   0   0
#>    0.172622359  0.410311789   0.123319040   0   0
#>    0.546196008  0.303980405   0.492525123   0   0
#>    0.463608431  0.386775403   0.585947427   0   0
#>    0.404748204  0.466417957   0.349340944   0   0
#>    0.194279403  0.410311789   0.123319040   0   0
#>    0.451043919  0.179254565   0.442290763   0   0
#>    0.555355499  0.168026766   0.451043919   0   0
#>    0.150788970  0.026238543   0.185353251   0   0
#>    0.207634981  0.178987717   0.223644958   0   0
#>    0.231474364  0.026238543   0.185353251   0   0
#>    0.273858559 -0.001428372   0.185642678   0   0
#>    0.303112958  0.036391768   0.337751698   0   0
#>    0.473142874  0.477386843   0.494156608   0   0
#>    0.185642678 -1.641873614  -0.053658492   0   0
#>    0.462756087  0.131218537   0.604736298   0   0
#>    0.601333212  0.451590397   0.548807208   0   0
#>    0.548807208  0.432331665   0.466824983   0   0
#>    0.508763418  0.477386843   0.494156608   0   0
#>    0.429766131  0.542123241   0.548580477   0   0
#>    0.587384914  0.318257145   0.621290172   0   0
#>    0.123319040  0.408101629   0.039906250   0   0
#>    0.065292800  0.218222004  -0.298194949   0   0
#>    0.337751698 -1.641873614  -0.053658492   0   0
#>    0.461675363 -0.038716678   0.580328387   0   0
#>    0.620305890  0.164345748   0.655685617   0   0
#>    0.548580477  0.504870469   0.508763418   0   0
#>    0.411572174  0.466433275   0.473142874   0   0
#>    0.407932804 -0.133570244   0.596267956   0   0
#>    0.596267956 -0.899998732   0.836771716   0   0
#>    0.126070949 -3.709231785  -0.475246891   0   0
#>   -0.475246891 -1.641873614  -0.053658492   0   0
#>    0.177410713  0.511531523   0.066648783   0   0
#>    0.039906250  0.388521581  -0.006989384   0   0
#>    0.066648783  0.557556154   0.026445533   0   0
#>    0.519342819  0.546948051   0.557394221   0   0
#>    0.025458827  0.557556154   0.026445533   0   0
#>    0.621290172  0.321614599   0.629821279   0   0
#>    0.585947427  0.438264256   0.601333212   0   0
#>    0.625514455  0.478838195   0.654492660   0   0
#>    0.026445533  0.388521581  -0.006989384   0   0
#>    0.126734630  0.518811513   0.177410713   0   0
#>    0.604736298 -0.038716678   0.580328387   0   0
#>    0.583337921  0.636593520   0.637109017   0   0
#>    0.261034038  0.648049073   0.251748719   0   0
#>    0.672061185  0.310930674   0.761616264   0   0
#>    0.411749403  0.732140328   0.482310964   0   0
#>    0.399336395  0.635081317   0.261034038   0   0
#>    0.714210490  0.436684524   0.794304527   0   0
#>    0.713938276  0.313043785   0.765758263   0   0
#>    0.255852516  0.773167422   0.339435512   0   0
#>    0.655685617  0.068378875   0.706090276   0   0
#>    0.654492660  0.372677627   0.714210490   0   0
#>    0.064497806  0.583019697   0.050883599   0   0
#>    0.050883599  0.572339690   0.126734630   0   0
#>    0.339435512  0.788295016   0.295329172   0   0
#>    0.765758263  0.310930674   0.761616264   0   0
#>    0.761348860  0.535752406   0.791413710   0   0
#>    0.679298017  0.180231386   0.780069209   0   0
#>    0.794015123  0.180231386   0.780069209   0   0
#>    0.168520042  0.648049073   0.251748719   0   0
#>    0.528707425  0.662381090   0.571971297   0   0
#>    0.400732686  0.772350740   0.345827400   0   0
#>    0.295329172  0.772545412   0.282926939   0   0
#>    0.811098935  0.127026602   0.838859395   0   0
#>    0.796518622  0.315511410   0.806491606   0   0
#>    0.661347165  0.760095433   0.689721572   0   0
#>    0.571971297  0.599409132   0.519342819   0   0
#>    0.282926939  0.764960039   0.230894991   0   0
#>    0.230894991  0.855151144   0.237319156   0   0
#>    0.162577033  0.676967130   0.123265795   0   0
#>    0.680267742  0.478838195   0.654492660   0   0
#>    0.123265795  0.572339690   0.126734630   0   0
#>    0.855680369  0.397316551   0.885989208   0   0
#>    0.409157460  0.874541273   0.409377650   0   0
#>    0.345827400  0.839292481   0.409157460   0   0
#>    0.265515959  0.788295016   0.295329172   0   0
#>    0.794314512  0.169632427   0.794015123   0   0
#>    0.794304527  0.417697278   0.832390768   0   0
#>    0.766173734  0.599852132   0.714916562   0   0
#>    0.121122655  0.883975792   0.144341079   0   0
#>    0.123708438  0.706642018   0.162577033   0   0
#>    0.010602336  0.804690434   0.123708438   0   0
#>   -0.007263701  0.663035115   0.064497806   0   0
#>    0.734334965  0.633222236   0.796700430   0   0
#>    0.797465937  0.068378875   0.706090276   0   0
#>    0.861786770  0.270310102   0.858723551   0   0
#>   -1.076295963  0.900291869  -3.733741832   0   0
#>    0.858723551  0.169632427   0.794015123   0   0
#>    0.855851919  0.535752406   0.791413710   0   0
#>    0.872577376  0.169358728   0.794314512   0   0
#>    0.832390768  0.539665571   0.855851919   0   0
#>    0.298382040  1.024669466   0.291251244   0   0
#>    0.041895068  0.900291869  -3.733741832   0   0
#>    0.437643507  0.874541273   0.409377650   0   0
#>    0.831572406  0.633222236   0.796700430   0   0
#>    0.796700430  0.604843982   0.766173734   0   0
#>    0.574854409  0.916680462   0.437643507   0   0
#>    0.836771716 -2.849193914   1.166183029   0   1
#>    0.838859395  0.165579606   0.797465937   0   0
#>    0.957360934  0.246524797   0.872577376   0   0
#>    1.038847864 -0.387335696   4.921154476   0   1
#>    0.929821292  0.268723130   0.861786770   0   0
#>    0.987752185  0.559223335   1.593494577   0   0
#>    0.144341079  0.885254229   0.144110934   0   0
#>    0.581383050  0.818160761   0.574854409   0   0
#>    0.875802945  0.656671375   1.042403714   0   0
#>    0.872405503  0.456284752   0.987752185   0   0
#>    0.867968022  0.693650640   0.910239335   0   0
#>    0.144110934  1.024669466   0.291251244   0   0
#>    0.133039846  6.064719379  -2.014937320   0   1
#>  -21.296046207  0.900291869  -3.733741832   0   0
#>    0.451849126  1.426432396   0.457037432   0   0
#>    0.291251244  1.336882079   0.451849126   0   0
#>    1.042403714  0.664322501   1.162519664   0   0
#>    0.910239335  0.810020326   0.904162999   0   0
#>    0.758705914  0.778991473   0.734334965   0   0
#>    0.957159293  0.810020326   0.904162999   0   0
#>    0.704576259  0.927061002   0.613814917   0   0
#>    0.817859899  0.697885676   0.867968022   0   0
#>    0.613814917  0.818160761   0.574854409   0   0
#>    0.471744736  0.927061002   0.613814917   0   0
#>    0.738471491  0.907659785   0.704576259   0   0
#>    0.737812961  0.949443061   0.716906725   0   0
#>    0.716906725  0.907659785   0.704576259   0   0
#>    0.866716557  1.110009439   0.868323229   0   0
#>    0.793963692  0.964966598   0.866716557   0   0
#>    0.765111544  0.942426689   0.793963692   0   0
#>    0.852662038  0.806931607   0.817859899   0   0
#>    0.959052680  0.923960269   0.882832973   0   0
#>    0.868323229  5.106416930   1.395975283   0   0
#>    0.979097090  0.800133900   0.957159293   0   0
#>    0.316507950  0.233126774   0.274381371   0   0
#>    0.221718972  0.330257246   0.172622359   0   0
#>    0.223734155  0.331372757   0.221718972   0   0
#>    0.423199271  0.303980405   0.492525123   0   0
#>    0.260597448  0.234203552   0.263626251   0   0
#>    0.463076212  0.421299962   0.463608431   0   0
#>    0.251074040  0.332955917   0.223734155   0   0
#>    0.239964253  0.098148791   0.294814105   0   0
#>    0.294814105  0.097698701   0.273858559   0   0
#>    0.092794776  0.236997212   0.260597448   0   0
#>    0.078738962  0.223625143   0.092794776   0   0
#>    0.316886184  0.095964356   0.303112958   0   0
#>    0.417278059  0.274999567   0.423199271   0   0
#>    0.352954288  0.411048393   0.406002499   0   0
#>    0.260754906  0.466417957   0.349340944   0   0
#>    0.442290763  0.248825633   0.417278059   0   0
#>    0.546196008  0.335346152   0.587384914   0   0
#>    0.466824983  0.466433275   0.473142874   0   0
#>    0.404748204  0.638380705   0.399336395   0   0
#>    0.194279403  0.518811513   0.177410713   0   0
#>    0.555355499  0.308155591   0.546196008   0   0
#>    0.150788970  0.131530714   0.126070949   0   0
#>    0.207634981  0.104233141   0.231474364   0   0
#>    0.273858559  0.104233141   0.231474364   0   0
#>    0.185642678  0.026238543   0.185353251   0   0
#>    0.462756087  0.122419267   0.461675363   0   0
#>    0.601333212  0.476123188   0.625514455   0   0
#>    0.548807208  0.477386843   0.494156608   0   0
#>    0.508763418  0.535731033   0.429766131   0   0
#>    0.461675363  0.063863397   0.407932804   0   0
#>    0.620305890  0.131218537   0.604736298   0   0
#>    0.411572174  0.541553901   0.404748204   0   0
#>    0.407932804  0.036391768   0.337751698   0   0
#>   -0.298194949 -0.213159274  -2.515031051   0   1
#>    0.126070949  0.204073553   0.065292800   0   0
#>   -0.475246891 -6.512646118  -1.023235163   0   1
#>    0.214505564  0.510579879   0.194279403   0   0
#>    0.039906250  0.511531523   0.066648783   0   0
#>    0.519342819  0.644500142   0.411749403   0   0
#>    0.025458827  0.714058466  -1.076295963   0   0
#>    0.621290172  0.235516366   0.672061185   0   0
#>    0.585947427  0.335346152   0.587384914   0   0
#>   -0.006989384  0.218222004  -0.298194949   0   0
#>    0.629821279  0.371296179   0.713938276   0   0
#>    0.583337921  0.476123188   0.625514455   0   0
#>    0.557394221  0.537925552   0.583337921   0   0
#>    0.580328387 -0.133570244   0.596267956   0   0
#>    0.672061185  0.189703400   0.679298017   0   0
#>    0.399336395  0.639733504   0.400732686   0   0
#>    0.714210490  0.371296179   0.713938276   0   0
#>    0.255852516  0.648049073   0.251748719   0   0
#>    0.761616264  0.321614599   0.629821279   0   0
#>    0.637109017  0.662381090   0.571971297   0   0
#>    0.050883599  0.559496617   0.025458827   0   0
#>    0.339435512  0.772350740   0.345827400   0   0
#>    0.761348860  0.436684524   0.794304527   0   0
#>    0.679298017  0.164345748   0.655685617   0   0
#>    0.168520042  0.706642018   0.162577033   0   0
#>    0.528707425  0.732140328   0.482310964   0   0
#>    0.400732686  0.644500142   0.411749403   0   0
#>    0.811098935 -0.899998732   0.836771716   0   0
#>    0.796518622  0.180231386   0.780069209   0   0
#>    0.661347165  0.599852132   0.714916562   0   0
#>    0.680267742  0.599852132   0.714916562   0   0
#>    0.123265795  0.663035115   0.064497806   0   0
#>    0.855680369  0.315511410   0.806491606   0   0
#>    0.409157460  0.732140328   0.482310964   0   0
#>    0.265515959  0.855151144   0.237319156   0   0
#>    0.121122655  0.804690434   0.123708438   0   0
#>    0.010602336  0.698422347  -0.007263701   0   0
#>   -0.007263701  0.714058466  -1.076295963   0   0
#>    0.734334965  0.760095433   0.689721572   0   0
#>    0.797465937  0.169358728   0.794314512   0   0
#>    0.237319156  0.883975792   0.144341079   0   0
#>    0.858723551  0.315511410   0.806491606   0   0
#>    0.855851919  0.551197765   0.872405503   0   0
#>    0.791413710  0.604843982   0.766173734   0   0
#>    0.872577376  0.268723130   0.861786770   0   0
#>    0.885989208  0.456284752   0.987752185   0   0
#>    0.832390768  0.396360418   0.855680369   0   0
#>    0.298382040  0.874541273   0.409377650   0   0
#>    0.041895068  0.835753934   0.121122655   0   0
#>    0.437643507  1.185169592   0.471744736   0   0
#>    0.831572406  0.697885676   0.867968022   0   0
#>    0.689721572  0.791263824   0.581383050   0   0
#>    0.838859395  0.181291749   0.957360934   0   0
#>    0.957360934  0.196261968   1.038847864   0   0
#>    0.929821292  0.397316551   0.885989208   0   0
#>    0.875802945  0.551197765   0.872405503   0   0
#>    0.133039846  0.885254229   0.144110934   0   0
#>  -21.296046207  2.008709531 -23.617196273   0   1
#>    0.451849126  1.185169592   0.471744736   0   0
#>    0.910239335  0.656671375   1.042403714   0   0
#>    0.758705914  0.861608856   0.738471491   0   0
#>    0.817859899  0.803863818   0.758705914   0   0
#>    0.738471491  0.942426689   0.793963692   0   0
#>    0.904162999  0.824936907   0.852662038   0   0
#>    0.737812961 27.047762880   2.852839839   0   1
#>    0.716906725  1.426432396   0.457037432   0   0
#>    0.457037432  3.304234780   0.466696043   0   1
#>    0.866716557  0.923960269   0.882832973   0   0
#>    0.765111544  0.966964551   0.737812961   0   0
#>    1.162519664  0.800133900   0.957159293   0   0
#>    0.852662038  0.923960269   0.882832973   0   0
#>    0.959052680  5.106416930   1.395975283   0   0
#>    1.593494577  0.664322501   1.162519664   0   0
#>    0.979097090  0.922117186   0.959052680   0   0
#>    1.395975283  9.484414817   1.952324004   0   1
#>    0.358414212  0.308432023   0.396368635   0   0
#>    0.322440340  0.460054594   0.352954288   0   0
#>    0.316714470  0.126283841   0.316886184   0   0
#>    0.463076212  0.411048393   0.406002499   0   0
#>    0.251074040  0.474863927   0.260754906   0   0
#>    0.078738962  0.178987717   0.223644958   0   0
#>    0.451043919  0.127283132   0.462756087   0   0
#>    0.555355499  0.150961090   0.620305890   0   0
#>    0.150788970  0.118147656   0.207634981   0   0
#>    0.429766131  0.537551287   0.411572174   0   0
#>    0.548580477  0.546948051   0.557394221   0   0
#>    0.214505564  0.635081317   0.261034038   0   0
#>    0.255852516  0.772545412   0.282926939   0   0
#>    0.637109017  0.647056908   0.661347165   0   0
#>    0.765758263  0.309374563   0.796518622   0   0
#>    0.761348860  0.489770312   0.680267742   0   0
#>    0.168520042  0.764960039   0.230894991   0   0
#>    0.528707425  0.791263824   0.581383050   0   0
#>    0.811098935  0.068378875   0.706090276   0   0
#>    0.265515959  0.874850291   0.298382040   0   0
#>    0.010602336  0.826780850   0.041895068   0   0
#>    0.831572406  0.596272887   0.875802945   0   0
#>    1.038847864  0.298484352   0.929821292   0   0
#>    0.133039846  1.881777704 -21.296046207   0   0
#>    0.765111544  1.110009439   0.868323229   0   0
#>    1.593494577  0.433360291   3.480454935   0   1
#>    0.979097090  2.420996174  10.562981460   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.29284031 0.28646461
#>   [2,] 0.73503347 0.90076192
#>   [3,] 0.60853147 0.09587632
#>   [4,] 0.62197573 0.01685974
#>   [5,] 0.33618124 0.96480794
#>   [6,] 0.26922383 0.32346001
#>   [7,] 0.75170758 0.25090001
#>   [8,] 0.04065085 0.47077173
#>   [9,] 0.43566510 0.51345839
#>  [10,] 0.09518253 0.74628051
#>  [11,] 0.30280813 0.70685540
#>  [12,] 0.00000000 0.42815704
#>  [13,] 0.89978632 0.54353396
#>  [14,] 0.93133180 0.02348915
#>  [15,] 0.69494257 0.60295763
#>  [16,] 0.65469759 0.52768398
#>  [17,] 0.39370191 0.30828270
#>  [18,] 0.05755845 0.14590559
#>  [19,] 0.49609981 0.87214054
#>  [20,] 0.33907371 0.09516107
#>  [21,] 0.57004062 0.82062904
#>  [22,] 0.86139441 0.65084490
#>  [23,] 0.77413074 0.01908522
#>  [24,] 0.61049590 0.15512799
#>  [25,] 0.37125115 0.75775995
#>  [26,] 0.40339940 0.19960766
#>  [27,] 0.06716017 0.08457707
#>  [28,] 0.08429426 0.92694368
#>  [29,] 1.00000000 0.18857535
#>  [30,] 0.49592643 0.02967131
#>  [31,] 0.94521163 0.76472111
#>  [32,] 0.23497463 0.59630813
#>  [33,] 0.91613595 0.80706641
#>  [34,] 0.99726093 0.72109866
#>  [35,] 0.66169000 0.73542987
#>  [36,] 0.54313422 0.74697582
#>  [37,] 0.22685618 0.51074065
#>  [38,] 0.45011069 0.51006492
#>  [39,] 0.78337233 0.24628383
#>  [40,] 0.17220296 0.27438668
#>  [41,] 0.82174595 0.48749968
#>  [42,] 0.26114436 0.89445399
#>  [43,] 0.61743321 0.57227221
#>  [44,] 0.65497046 0.89274065
#>  [45,] 0.20671113 0.20702213
#>  [46,] 0.48244126 0.52532576
#>  [47,] 0.97184938 0.67445584
#>  [48,] 0.92679433 0.33089188
#>  [49,] 0.50604875 0.48090678
#>  [50,] 0.96560657 0.98397542
#>  [51,] 0.07570866 0.19482753
#>  [52,] 0.24688819 0.83428633
#>  [53,] 0.50525146 0.82456352
#>  [54,] 0.29785683 0.87839438
#>  [55,] 0.56206674 0.92642143
#>  [56,] 0.86378672 1.00000000
#>  [57,] 0.54607596 0.08708244
#>  [58,] 0.04374852 0.24057473
#>  [59,] 0.92347183 0.26108568
#>  [60,] 0.47995877 0.09142809
#>  [61,] 0.05837684 0.22413707
#>  [62,] 0.78536762 0.21827123
#>  [63,] 0.22163200 0.86405291
#>  [64,] 0.38008978 0.35491849
#>  [65,] 0.69719660 0.81842084
#>  [66,] 0.07210092 0.65792558
#>  [67,] 0.32326468 0.29756133
#>  [68,] 0.12448565 0.37942849
#>  [69,] 0.16111061 0.19671328
#>  [70,] 0.72742187 0.31842528
#>  [71,] 0.52288707 0.48748473
#>  [72,] 0.23835248 0.84756675
#>  [73,] 0.32776754 0.70887873
#>  [74,] 0.82272604 0.33060215
#>  [75,] 0.75993996 0.39696889
#>  [76,] 0.08946118 0.53705827
#>  [77,] 0.41546951 0.72242299
#>  [78,] 0.17237344 0.35920075
#>  [79,] 0.57172184 0.65615089
#>  [80,] 0.72582444 0.06806990
#>  [81,] 0.87521737 0.94136436
#>  [82,] 0.54558254 0.32276330
#>  [83,] 0.96981713 0.94365189
#>  [84,] 0.97152019 0.78990755
#>  [85,] 0.73946283 0.30204453
#>  [86,] 0.25841619 0.20290179
#>  [87,] 0.73579548 0.91535529
#>  [88,] 0.38787906 0.67337537
#>  [89,] 0.25835098 0.74429276
#>  [90,] 0.38299380 0.49859433
#>  [91,] 0.76999873 0.03364192
#>  [92,] 0.99417186 0.75919053
#>  [93,] 0.50179539 0.00000000
#>  [94,] 0.46442086 0.43282396
#>  [95,] 0.11014682 0.73036227
#>  [96,] 0.88027893 0.02115191
#>  [97,] 0.15017002 0.23828897
#>  [98,] 0.16916287 0.53484910
#>  [99,] 0.55469345 0.48577354
#> [100,] 0.37112278 0.45043173
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,67,6), [2,7,5]
#> 2: (64,6,67), [1,3,11]
#> 3: (17,64,67), [2,13,22]
#> 4: (26,1,86), [10,26,5]
#> 5: (26,67,1), [1,4,13]
#> 6: (45,1,40), [7,14,10]
#> 7: (1,6,40), [8,6,1]
#> 8: (78,40,6), [7,21,19]
#> 9: (37,6,100), [11,20,21]
#> 10: (45,86,1), [4,6,16]
#> 11: (6,64,100), [23,9,2]
#> 12: (100,94,9), [29,28,23]
#> 13: (26,17,67), [3,5,24]
#> 14: (97,45,40), [6,15,18]
#> 15: (97,40,58), [39,38,14]
#> 16: (45,20,86), [26,10,17]
#> 17: (45,69,20), [50,16,18]
#> 18: (69,45,97), [14,36,17]
#> 19: (40,78,68), [25,39,8]
#> 20: (37,100,90), [28,27,9]
#> 21: (37,78,6), [8,9,25]
#> 22: (64,17,94), [31,23,3]
#> 23: (94,100,64), [11,22,12]
#> 24: (17,26,82), [63,31,13]
#> 25: (37,68,78), [19,21,33]
#> 26: (86,20,26), [49,4,16]
#> 27: (32,37,90), [20,48,34]
#> 28: (100,9,90), [70,20,12]
#> 29: (38,9,94), [12,40,44]
#> 30: (99,94,82), [31,83,56]
#> 31: (17,82,94), [30,22,24]
#> 32: (24,26,60), [49,64,63]
#> 33: (98,68,37), [25,34,42]
#> 34: (32,98,37), [33,27,54]
#> 35: (69,51,18), [51,61,36]
#> 36: (51,69,97), [18,37,35]
#> 37: (97,61,51), [51,36,38]
#> 38: (97,58,61), [41,37,15]
#> 39: (40,68,58), [52,15,19]
#> 40: (94,49,38), [45,29,56]
#> 41: (61,58,18), [57,51,38]
#> 42: (68,98,76), [75,53,33]
#> 43: (88,9,46), [44,71,70]
#> 44: (46,9,38), [29,45,43]
#> 45: (46,38,49), [40,46,44]
#> 46: (71,46,49), [45,47,55]
#> 47: (49,99,71), [55,46,56]
#> 48: (90,88,32), [69,27,70]
#> 49: (26,20,60), [65,32,26]
#> 50: (69,27,20), [60,17,61]
#> 51: (61,18,51), [35,37,41]
#> 52: (68,12,58), [57,39,58]
#> 53: (68,76,8), [80,58,42]
#> 54: (98,32,66), [87,75,34]
#> 55: (99,46,71), [46,47,78]
#> 56: (99,49,94), [40,30,47]
#> 57: (12,18,58), [41,52,62]
#> 58: (68,8,12), [59,52,53]
#> 59: (8,66,12), [147,58,80]
#> 60: (93,20,27), [50,0,73]
#> 61: (69,18,27), [62,50,35]
#> 62: (27,18,12), [57,0,61]
#> 63: (24,82,26), [24,32,79]
#> 64: (24,60,57), [66,74,32]
#> 65: (60,20,30), [73,66,49]
#> 66: (60,30,57), [72,64,65]
#> 67: (16,43,99), [78,82,107]
#> 68: (4,57,93), [72,132,92]
#> 69: (32,88,11), [76,81,48]
#> 70: (90,9,88), [43,48,28]
#> 71: (46,79,88), [89,43,77]
#> 72: (57,30,93), [73,68,66]
#> 73: (93,30,20), [65,60,72]
#> 74: (3,24,57), [64,92,112]
#> 75: (98,66,76), [80,42,54]
#> 76: (73,11,88), [69,85,88]
#> 77: (46,43,79), [90,71,78]
#> 78: (46,99,43), [67,77,55]
#> 79: (82,24,70), [94,83,63]
#> 80: (8,76,66), [75,59,53]
#> 81: (32,11,89), [88,97,69]
#> 82: (99,75,16), [127,67,101]
#> 83: (99,82,70), [79,101,30]
#> 84: (88,77,25), [119,85,89]
#> 85: (88,25,73), [95,76,84]
#> 86: (7,85,70), [93,94,108]
#> 87: (32,95,66), [104,54,97]
#> 88: (73,89,11), [81,76,95]
#> 89: (79,77,88), [84,71,111]
#> 90: (15,79,43), [77,107,106]
#> 91: (80,3,4), [92,125,112]
#> 92: (4,3,57), [74,68,91]
#> 93: (70,85,74), [102,116,86]
#> 94: (7,70,24), [79,99,86]
#> 95: (25,89,73), [88,85,105]
#> 96: (77,36,53), [135,119,111]
#> 97: (32,89,95), [113,87,81]
#> 98: (72,95,52), [113,133,118]
#> 99: (62,7,24), [94,110,109]
#> 100: (41,15,16), [107,127,155]
#> 101: (99,70,75), [116,82,83]
#> 102: (74,85,39), [108,117,93]
#> 103: (66,10,28), [148,147,104]
#> 104: (10,66,95), [87,128,103]
#> 105: (89,25,52), [120,113,95]
#> 106: (79,15,35), [145,130,90]
#> 107: (15,43,16), [67,100,90]
#> 108: (85,7,39), [109,102,86]
#> 109: (7,62,39), [131,108,99]
#> 110: (62,24,80), [112,123,99]
#> 111: (36,77,79), [89,130,96]
#> 112: (80,24,3), [74,91,110]
#> 113: (52,95,89), [97,105,98]
#> 114: (25,19,54), [137,120,138]
#> 115: (75,74,41), [142,127,116]
#> 116: (70,74,75), [115,101,93]
#> 117: (59,74,39), [102,131,139]
#> 118: (63,95,72), [98,136,128]
#> 119: (77,53,25), [138,84,96]
#> 120: (54,52,25), [105,114,133]
#> 121: (21,36,35), [130,144,135]
#> 122: (91,96,62), [153,123,140]
#> 123: (91,62,80), [110,124,122]
#> 124: (23,91,80), [123,125,140]
#> 125: (23,80,4), [91,132,124]
#> 126: (22,65,35), [144,145,166]
#> 127: (41,16,75), [82,115,100]
#> 128: (63,10,95), [104,118,148]
#> 129: (42,72,54), [133,151,136]
#> 130: (79,35,36), [121,111,106]
#> 131: (59,39,62), [109,153,117]
#> 132: (4,93,23), [154,125,68]
#> 133: (54,72,52), [98,120,129]
#> 134: (19,53,21), [135,157,138]
#> 135: (21,53,36), [96,121,134]
#> 136: (42,63,72), [118,129,149]
#> 137: (5,54,19), [114,152,151]
#> 138: (25,53,19), [134,114,119]
#> 139: (74,59,48), [163,142,117]
#> 140: (91,23,96), [154,122,124]
#> 141: (13,41,48), [142,171,146]
#> 142: (74,48,41), [141,115,139]
#> 143: (44,21,65), [144,158,156]
#> 144: (65,21,35), [121,126,143]
#> 145: (22,35,15), [106,155,126]
#> 146: (22,41,13), [141,170,155]
#> 147: (66,28,12), [0,59,103]
#> 148: (28,10,63), [128,149,103]
#> 149: (28,63,42), [136,150,148]
#> 150: (42,5,28), [0,149,151]
#> 151: (5,42,54), [129,137,150]
#> 152: (19,55,5), [183,137,157]
#> 153: (62,96,59), [159,131,122]
#> 154: (96,23,93), [132,161,140]
#> 155: (15,41,22), [146,145,100]
#> 156: (21,44,55), [164,157,143]
#> 157: (21,55,19), [152,134,156]
#> 158: (65,2,44), [165,143,169]
#> 159: (96,29,59), [163,153,160]
#> 160: (96,14,29), [0,159,161]
#> 161: (14,96,93), [154,0,160]
#> 162: (48,29,47), [176,171,163]
#> 163: (59,29,48), [162,139,159]
#> 164: (44,87,55), [180,156,165]
#> 165: (44,2,87), [173,164,158]
#> 166: (33,65,22), [126,172,169]
#> 167: (56,87,81), [173,186,180]
#> 168: (31,22,47), [170,175,172]
#> 169: (33,2,65), [158,166,181]
#> 170: (47,22,13), [146,171,168]
#> 171: (48,47,13), [170,141,162]
#> 172: (33,22,31), [168,178,166]
#> 173: (81,87,2), [165,181,167]
#> 174: (92,31,34), [175,0,179]
#> 175: (34,31,47), [168,176,174]
#> 176: (34,47,29), [162,0,175]
#> 177: (33,84,83), [184,185,178]
#> 178: (31,84,33), [177,172,179]
#> 179: (92,84,31), [178,174,184]
#> 180: (56,55,87), [164,167,183]
#> 181: (81,2,33), [169,185,173]
#> 182: (50,81,83), [185,187,186]
#> 183: (56,5,55), [152,180,0]
#> 184: (84,92,83), [187,177,179]
#> 185: (33,83,81), [182,181,177]
#> 186: (50,56,81), [167,182,0]
#> 187: (50,83,92), [184,0,182]
#> boundary nodes:  27 93 14 29 34 92 50 56 5 28 12 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.05992625 0.08073993 0.09037569 0.03238487 0.12124045 0.10872553 0.16280421 
#>                                                                              
#> 0.09593324 0.08560844 0.07124852 0.07568882 0.13282994 0.13458714 0.20489928 
#>                                                                              
#> 0.06461047 0.05198463 0.04960401 0.10323468 0.14321905 0.18810465 0.12266575 
#>                                                                              
#> 0.15660935 0.05472848 0.13675001 0.13252984 0.16650108 0.06252788 0.05218033 
#>                                                                              
#> 0.04299110 0.03405056 0.02200419 0.16062021 0.06308141 0.09567110 0.07973230 
#>                                                                              
#> 0.04825809 0.01483882 0.05030268 0.03185238 0.17122829 0.14093451 0.27211945 
#>                                                                              
#> 0.19261635 0.08231861 0.17413948 0.05538779 0.06359905 0.05889400 0.24081555 
#>                                                                              
#> 0.06207561 0.28806050 0.06625985 0.16997554 0.07623000 0.10685643 0.09769401 
#>                                                                              
#> 0.09142196 0.16690401 0.18464790 0.03024618 0.09370233 0.12210768 0.09552572 
#>                                                                              
#> 0.20084244 0.05812052 0.16781655 0.18189107 0.05660356 0.06542323 0.02033012 
#>                                                                              
#> 0.08182031 0.05627523 0.08015216 0.06307157 0.08802443 0.07792825 0.08634208 
#>                                                                              
#> 0.14885742 0.17175102 0.17062537 0.08535678 0.08500896 0.07097393 0.18099098 
#>                                                                              
#> 0.14602203 0.13658195 0.05810976 0.03199946 0.02808358 0.14449838 0.16972560 
#>                                                                              
#> 0.05928422 0.19834160 0.15690059 0.09135993 0.09304995 0.17370545 0.14972909 
#>                                                                              
#> 0.11911667 0.21877885 0.16157298 0.05600584 0.11578876 0.09026749 0.02184767 
#>                                                                              
#> 0.06701575 0.27300327 0.01578695 0.06490851 0.02348123 0.04844921 0.06988523 
#>                                                                              
#> 0.10616831 0.18857668 0.12717515 0.12513396 0.09607362 0.50585930 0.17279990 
#>                                                                              
#> 0.04985431 0.25471701 0.04007139 0.22912395 0.24379057 0.16808826 0.09882053 
#>                                                                              
#> 0.08542863 0.08046382 0.10542442 0.17879814 0.37907413 0.48669530 0.16004083 
#>                                                                              
#> 0.17408082 0.01461325 0.23461163 0.14182714 0.11295033 0.09061834 0.11397182 
#>                                                                              
#> 0.14944438 0.14139702 0.06791206 0.09411370 0.15375377 0.05798139 0.03642111 
#>                                                                              
#> 0.20390428 0.09462742 0.18606221 0.05973942 0.04389075 0.14502063 0.14059673 
#>                                                                              
#> 0.15636501 0.11717228 0.09021126 0.12655629 0.04229061 0.10644605 0.05186897 
#>                                                                              
#> 0.17329959 0.11537211 0.19201316 0.11480851 0.15256937 0.16103636 0.17736307 
#>                                                                              
#> 0.07855538 0.16326465 0.14525020 0.08595175 0.11534501 0.08621712 0.09287769 
#>                                                                              
#> 0.07823575 0.16147403 0.17567213 0.03575135 0.01807756 0.12398331 0.15674451 
#>                                                                              
#> 0.10465562 0.13368301 0.44278782 0.14629154 0.35008240 0.21181927 0.06378766 
#>                                                                              
#> 0.10843114 0.12135718 0.12969997 0.17484930 0.18850457 0.06981320 0.15838636 
#>                                                                              
#> 0.14292235 0.18977796 0.14981957 0.24050062 0.08600723 0.07175972 0.02504128 
#>                                                                              
#> 0.08336243 0.10340214 0.09607893 0.13599076 0.18317340 0.18592252 0.17181366 
#>                                                                              
#> 0.22363451 0.26929429 0.09072067 0.11991424 0.09521774 0.12054393 0.14120692 
#>                                                                              
#> 0.10961662 0.14087926 0.18526787 0.06879678 0.15217129 0.21687893 0.17407508 
#>                                                                              
#> 0.14458856 0.06740417 0.09011481 0.07841398 0.05213332 0.18482747 0.16932249 
#>                                                                              
#> 0.10406869 0.11098524 0.21435038 0.08547796 0.17320337 0.15105265 0.17980973 
#>                                                                              
#> 0.09453075 0.10609247 0.20582454 0.43017819 0.34650563 0.08392916 0.16553488 
#>                                                                              
#> 0.21923358 0.05136652 0.14594548 0.03821693 0.05311590 0.53253036 0.14675580 
#>                                                                              
#> 0.04927161 0.15344864 0.14039327 0.04054276 0.31056199 0.09992952 0.22659263 
#>                                                                              
#> 0.11524276 0.04858176 0.08481424 0.09494507 0.10910685 0.04675126 0.16171463 
#>                                                                              
#> 0.09004666 0.08542277 0.04888755 0.08236964 0.17976469 0.05258988 0.13421204 
#>                                                                              
#> 0.11370055 0.13000431 0.04687895 0.17149238 0.09132006 0.12240202 0.01513179 
#>                                                                   
#> 0.11141437 0.10286015 0.05110635 0.03816586 0.52877786 0.10307313 
#> 
#> $alpha
#> [1] 0.3058544
#> 
#> attr(,"class")
#> [1] "scree"