Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1           y1         x2           y2         mx1
#>    49   65 0.23826185 0.9713806455 0.04732251 0.9439205380  0.13435178
#>    76   90 0.11605180 0.8505144414 0.26054264 0.9287583050  0.17627298
#>     1   76 0.08025720 0.9168790837 0.11605180 0.8505144414  0.16725015
#>     1   65 0.08025720 0.9168790837 0.04732251 0.9439205380 -0.03235053
#>    17   71 0.06122162 0.6125345127 0.16304251 0.7644512574  0.02660541
#>    59   23 0.36233749 0.8907970183 0.40405074 1.0000000000  0.33201422
#>    22   59 0.32728318 0.7452388683 0.36233749 0.8907970183  0.28279098
#>    23   49 0.40405074 1.0000000000 0.23826185 0.9713806455  0.32085727
#>    71   22 0.16304251 0.7644512574 0.32728318 0.7452388683  0.25303388
#>    18   23 0.47185342 0.9769194759 0.40405074 1.0000000000  0.41867961
#>    59   56 0.36233749 0.8907970183 0.35457890 0.7406075735  0.40444162
#>    22   52 0.32728318 0.7452388683 0.25178173 0.6574059527  0.30260897
#>    18   59 0.47185342 0.9769194759 0.36233749 0.8907970183  0.42990418
#>    71   52 0.16304251 0.7644512574 0.25178173 0.6574059527  0.24352786
#>    56   94 0.35457890 0.7406075735 0.35917950 0.7113160521  0.47216653
#>    56   22 0.35457890 0.7406075735 0.32728318 0.7452388683  0.33753775
#>    18   55 0.47185342 0.9769194759 0.47751943 0.8626626106  0.47603003
#>     6   98 0.56645419 0.8854680274 0.53198354 0.9318032775  0.51890491
#>    17    8 0.06122162 0.6125345127 0.20461746 0.5795403111  0.14857297
#>    65   17 0.04732251 0.9439205380 0.06122162 0.6125345127 -0.08814326
#>    52    8 0.25178173 0.6574059527 0.20461746 0.5795403111  0.24875505
#>    55   56 0.47751943 0.8626626106 0.35457890 0.7406075735  0.47245781
#>    53   57 0.35765478 0.5258159409 0.44928869 0.5344208849  0.39525008
#>    94   52 0.35917950 0.7113160521 0.25178173 0.6574059527  0.31869952
#>    82    8 0.18637549 0.5078898838 0.20461746 0.5795403111  0.12500976
#>    78   82 0.05210580 0.5189323421 0.18637549 0.5078898838  0.12301227
#>    82   19 0.18637549 0.5078898838 0.29883329 0.5941228671  0.26097929
#>    57   94 0.44928869 0.5344208849 0.35917950 0.7113160521  0.39555913
#>    27    6 0.55645215 0.6620035135 0.56645419 0.8854680274  0.54755465
#>    27   94 0.55645215 0.6620035135 0.35917950 0.7113160521  0.45317681
#>    33   18 0.76090906 0.9753487642 0.47185342 0.9769194759  0.61709708
#>    37   33 0.69401073 0.9186485788 0.76090906 0.9753487642  0.62461630
#>    37    6 0.69401073 0.9186485788 0.56645419 0.8854680274  0.62820614
#>    46   28 0.71394613 0.7602742659 0.68682803 0.8786845954  0.62210327
#>    11   28 0.73987745 0.8719858348 0.68682803 0.8786845954  0.70647569
#>    80   94 0.54974355 0.6309822231 0.35917950 0.7113160521  0.44227790
#>    66   57 0.38725524 0.4628398578 0.44928869 0.5344208849  0.40537944
#>    66   40 0.38725524 0.4628398578 0.50076329 0.5137606018  0.45279777
#>    82    2 0.18637549 0.5078898838 0.32761797 0.4311551574  0.27374272
#>    82    7 0.18637549 0.5078898838 0.22097551 0.4102981086  0.25449218
#>    78   16 0.05210580 0.5189323421 0.11077601 0.3887808520  0.11560956
#>     2   66 0.32761797 0.4311551574 0.38725524 0.4628398578  0.34031616
#>    80   40 0.54974355 0.6309822231 0.50076329 0.5137606018  0.55209715
#>    47   40 0.59626616 0.5070209645 0.50076329 0.5137606018  0.54402753
#>    37   28 0.69401073 0.9186485788 0.68682803 0.8786845954  0.71572458
#>    27   83 0.55645215 0.6620035135 0.66617148 0.6834855716  0.59180463
#>    79   74 0.28601278 0.4093335511 0.18718042 0.3245232712  0.26362445
#>    83   46 0.66617148 0.6834855716 0.71394613 0.7602742659  0.60133297
#>    33   23 0.76090906 0.9753487642 0.40405074 1.0000000000  0.61934378
#>    69   33 0.77090942 0.8868418531 0.76090906 0.9753487642  0.74312121
#>     7   93 0.22097551 0.4102981086 0.15552104 0.3316881077  0.18219688
#>    16    7 0.11077601 0.3887808520 0.22097551 0.4102981086  0.15744750
#>    27   80 0.55645215 0.6620035135 0.54974355 0.6309822231  0.61925275
#>    11   69 0.73987745 0.8719858348 0.77090942 0.8868418531  0.73791281
#>    11   46 0.73987745 0.8719858348 0.71394613 0.7602742659  0.77014919
#>    16   99 0.11077601 0.3887808520 0.03587036 0.3664911153  0.08306412
#>    78   65 0.05210580 0.5189323421 0.04732251 0.9439205380 -2.40722712
#>    83   80 0.66617148 0.6834855716 0.54974355 0.6309822231  0.62052478
#>    80   47 0.54974355 0.6309822231 0.59626616 0.5070209645  0.59933118
#>    79    7 0.28601278 0.4093335511 0.22097551 0.4102981086  0.25258182
#>    12   25 0.67124912 0.5799729013 0.71568534 0.6348133680  0.66373433
#>    25   46 0.71568534 0.6348133680 0.71394613 0.7602742659  0.72886152
#>    74   84 0.18718042 0.3245232712 0.17634711 0.3023809243  0.20651214
#>    64  100 0.82917869 0.6667617476 0.81784737 0.7514760015  0.76123204
#>    69   51 0.77090942 0.8868418531 0.96650476 0.9310825127  0.87099771
#>     2   79 0.32761797 0.4311551574 0.28601278 0.4093335511  0.36883862
#>    74   93 0.18718042 0.3245232712 0.15552104 0.3316881077  0.16940309
#>     3   84 0.10835562 0.2995831712 0.17634711 0.3023809243  0.14237958
#>    93   16 0.15552104 0.3316881077 0.11077601 0.3887808520  0.11256250
#>    75   73 0.54321515 0.3670931795 0.57361494 0.3256280762  0.61699847
#>     5   15 0.26678220 0.1932275582 0.47699423 0.2202285460  0.36070694
#>    86   13 0.11962118 0.2103560328 0.21773799 0.2733261911  0.16679975
#>    38   12 0.83900952 0.4979888648 0.67124912 0.5799729013  0.72617975
#>    47   75 0.59626616 0.5070209645 0.54321515 0.3670931795  0.63211464
#>    64   24 0.82917869 0.6667617476 0.87020438 0.7138607743  0.82748912
#>    25   64 0.71568534 0.6348133680 0.82917869 0.6667617476  0.75909733
#>   100   51 0.81784737 0.7514760015 0.96650476 0.9310825127  0.88213641
#>    77   24 0.85370972 0.6674933541 0.87020438 0.7138607743  0.84051438
#>    26   77 0.81701689 0.6319234922 0.85370972 0.6674933541  0.84217322
#>    24   51 0.87020438 0.7138607743 0.96650476 0.9310825127  0.91091379
#>     3   86 0.10835562 0.2995831712 0.11962118 0.2103560328  0.14408833
#>     3   99 0.10835562 0.2995831712 0.03587036 0.3664911153 -0.01497774
#>    47   97 0.59626616 0.5070209645 0.62911235 0.3132129714  0.66689957
#>    75   66 0.54321515 0.3670931795 0.38725524 0.4628398578  0.41663670
#>    13    5 0.21773799 0.2733261911 0.26678220 0.1932275582  0.34356200
#>    51   45 0.96650476 0.9310825127 0.93693612 0.9791703724  0.87502438
#>    38   31 0.83900952 0.4979888648 0.92455179 0.5522007592  0.85383197
#>    31   77 0.92455179 0.5522007592 0.85370972 0.6674933541  0.87965121
#>    73   15 0.57361494 0.3256280762 0.47699423 0.2202285460  0.55043622
#>    13   67 0.21773799 0.2733261911 0.15284447 0.1469726250  0.19488134
#>     4   39 0.00000000 0.0498495959 0.04946479 0.0099598093  0.07798922
#>    31   95 0.92455179 0.5522007592 1.00000000 0.6302007247  0.92207289
#>    26   25 0.81701689 0.6319234922 0.71568534 0.6348133680  0.76410076
#>    85   15 0.47344819 0.1653212179 0.47699423 0.2202285460  0.41889557
#>    12   47 0.67124912 0.5799729013 0.59626616 0.5070209645  0.70546712
#>    61   92 0.62761360 0.2676992779 0.63749827 0.2940841139  0.60496969
#>    73   61 0.57361494 0.3256280762 0.62761360 0.2676992779  0.59512895
#>    97   92 0.62911235 0.3132129714 0.63749827 0.2940841139  0.70361011
#>    72   41 0.23259280 0.1440834440 0.40908648 0.1343928417  0.31977753
#>    95   24 1.00000000 0.6302007247 0.87020438 0.7138607743  0.93123569
#>    10   42 0.93920507 0.4176845935 0.93449772 0.4895606211  0.88289771
#>    85   30 0.47344819 0.1653212179 0.47963029 0.1401318697  0.56362185
#>    61   15 0.62761360 0.2676992779 0.47699423 0.2202285460  0.56385230
#>    85   41 0.47344819 0.1653212179 0.40908648 0.1343928417  0.44375489
#>    31   42 0.92455179 0.5522007592 0.93449772 0.4895606211  0.95677110
#>    62   60 0.17355531 0.0613715131 0.19375275 0.0556529365  0.19733307
#>    67   62 0.15284447 0.1469726250 0.17355531 0.0613715131  0.19146814
#>    99    4 0.03587036 0.3664911153 0.00000000 0.0498495959 -0.05620989
#>    96   10 0.90269672 0.3325988524 0.93920507 0.4176845935  0.85244905
#>    44   32 0.29827744 0.0196741388 0.50302728 0.0160756123  0.40089066
#>    61   58 0.62761360 0.2676992779 0.65112311 0.1975905993  0.68130365
#>    20   36 0.73691411 0.2315159545 0.80675176 0.3223175021  0.73140554
#>    39   62 0.04946479 0.0099598093 0.17355531 0.0613715131  0.09042251
#>    67   86 0.15284447 0.1469726250 0.11962118 0.2103560328  0.05311176
#>    99   65 0.03587036 0.3664911153 0.04732251 0.9439205380 -4.77072919
#>    34   92 0.77010333 0.3742238984 0.63749827 0.2940841139  0.71767013
#>    63   42 0.98387039 0.5573375645 0.93449772 0.4895606211  1.04528998
#>    34   10 0.77010333 0.3742238984 0.93920507 0.4176845935  0.85097706
#>    36   96 0.80675176 0.3223175021 0.90269672 0.3325988524  0.84798771
#>    39   68 0.04946479 0.0099598093 0.15852706 0.0000000000  0.10764922
#>    68   60 0.15852706 0.0000000000 0.19375275 0.0556529365  0.17514428
#>    58   30 0.65112311 0.1975905993 0.47963029 0.1401318697  0.57720142
#>    32   30 0.50302728 0.0160756123 0.47963029 0.1401318697  0.44980824
#>    72   60 0.23259280 0.1440834440 0.19375275 0.0556529365  0.26030165
#>    44   60 0.29827744 0.0196741388 0.19375275 0.0556529365  0.23073448
#>    43   36 0.89433514 0.2589198946 0.80675176 0.3223175021  0.81602308
#>    88   20 0.74369286 0.1621239967 0.73691411 0.2315159545  0.70249176
#>    43   20 0.89433514 0.2589198946 0.73691411 0.2315159545  0.81743962
#>    35   88 0.72829629 0.1360126205 0.74369286 0.1621239967  0.69489637
#>    81   35 0.70681627 0.1208828962 0.72829629 0.1360126205  0.69015138
#>    14   32 0.57273723 0.0001042156 0.50302728 0.0160756123  0.38906456
#>    21   10 0.93136895 0.2961744298 0.93920507 0.4176845935  0.96836797
#>    43   21 0.89433514 0.2589198946 0.93136895 0.2961744298  0.89401776
#>    50   20 0.89797532 0.2317769092 0.73691411 0.2315159545  0.81745534
#>    21   29 0.93136895 0.2961744298 0.97361139 0.2803804358  0.97714907
#>    29   63 0.97361139 0.2803804358 0.98387039 0.5573375645  1.20205224
#>    30   48 0.47963029 0.1401318697 0.58000578 0.0332018850  0.57209633
#>    81    9 0.70681627 0.1208828962 0.61586053 0.0437430849  0.68297736
#>    81   58 0.70681627 0.1208828962 0.65112311 0.1975905993  0.62768079
#>    50   29 0.89797532 0.2317769092 0.97361139 0.2803804358  0.93899655
#>     9   14 0.61586053 0.0437430849 0.57273723 0.0001042156  0.66901616
#>     9   48 0.61586053 0.0437430849 0.58000578 0.0332018850  0.60627897
#>    48   32 0.58000578 0.0332018850 0.50302728 0.0160756123  0.54159401
#>    68   14 0.15852706 0.0000000000 0.57273723 0.0001042156  0.36587537
#>    29   95 0.97361139 0.2803804358 1.00000000 0.6302007247  1.94350019
#>    35   70 0.72829629 0.1360126205 0.83433588 0.1460736440  0.78319535
#>    54   91 0.74461723 0.0272072620 0.84401445 0.0047221414  0.80726626
#>    50   70 0.89797532 0.2317769092 0.83433588 0.1460736440  0.94247464
#>    91   89 0.84401445 0.0047221414 0.88034794 0.0426704320  0.81025032
#>    14   91 0.57273723 0.0001042156 0.84401445 0.0047221414  0.71417578
#>    49   90 0.23826185 0.9713806455 0.26054264 0.9287583050  0.17627298
#>    49   76 0.23826185 0.9713806455 0.11605180 0.8505144414  0.16725015
#>    76   65 0.11605180 0.8505144414 0.04732251 0.9439205380 -0.03235053
#>    71   76 0.16304251 0.7644512574 0.11605180 0.8505144414  0.21150800
#>    90   23 0.26054264 0.9287583050 0.40405074 1.0000000000  0.33201422
#>    90   59 0.26054264 0.9287583050 0.36233749 0.8907970183  0.28279098
#>    90   22 0.26054264 0.9287583050 0.32728318 0.7452388683  0.25303388
#>    55   59 0.47751943 0.8626626106 0.36233749 0.8907970183  0.42990418
#>    94   22 0.35917950 0.7113160521 0.32728318 0.7452388683  0.33753775
#>    98   55 0.53198354 0.9318032775 0.47751943 0.8626626106  0.47603003
#>    71    8 0.16304251 0.7644512574 0.20461746 0.5795403111  0.14857297
#>    76   17 0.11605180 0.8505144414 0.06122162 0.6125345127 -0.08814326
#>    19    8 0.29883329 0.5941228671 0.20461746 0.5795403111  0.24875505
#>    27   56 0.55645215 0.6620035135 0.35457890 0.7406075735  0.47245781
#>    19   57 0.29883329 0.5941228671 0.44928869 0.5344208849  0.39525008
#>    19   52 0.29883329 0.5941228671 0.25178173 0.6574059527  0.31869952
#>    17   82 0.06122162 0.6125345127 0.18637549 0.5078898838  0.12301227
#>    19   94 0.29883329 0.5941228671 0.35917950 0.7113160521  0.39555913
#>    53   82 0.35765478 0.5258159409 0.18637549 0.5078898838  0.27252437
#>    55    6 0.47751943 0.8626626106 0.56645419 0.8854680274  0.54755465
#>    98   18 0.53198354 0.9318032775 0.47185342 0.9769194759  0.61709708
#>    98   33 0.53198354 0.9318032775 0.76090906 0.9753487642  0.62461630
#>    28    6 0.68682803 0.8786845954 0.56645419 0.8854680274  0.62820614
#>    57   40 0.44928869 0.5344208849 0.50076329 0.5137606018  0.45279777
#>    53    2 0.35765478 0.5258159409 0.32761797 0.4311551574  0.27374272
#>    82   16 0.18637549 0.5078898838 0.11077601 0.3887808520  0.11560956
#>    53   66 0.35765478 0.5258159409 0.38725524 0.4628398578  0.34031616
#>    57   80 0.44928869 0.5344208849 0.54974355 0.6309822231  0.49883495
#>    75   40 0.54321515 0.3670931795 0.50076329 0.5137606018  0.54402753
#>     6   83 0.56645419 0.8854680274 0.66617148 0.6834855716  0.59180463
#>    13   74 0.21773799 0.2733261911 0.18718042 0.3245232712  0.26362445
#>    79   82 0.28601278 0.4093335511 0.18637549 0.5078898838  0.26919398
#>     6   46 0.56645419 0.8854680274 0.71394613 0.7602742659  0.60133297
#>    37   69 0.69401073 0.9186485788 0.77090942 0.8868418531  0.73791281
#>   100   46 0.81784737 0.7514760015 0.71394613 0.7602742659  0.77014919
#>    99   78 0.03587036 0.3664911153 0.05210580 0.5189323421  0.05428487
#>    12   80 0.67124912 0.5799729013 0.54974355 0.6309822231  0.62052478
#>    74    7 0.18718042 0.3245232712 0.22097551 0.4102981086  0.25258182
#>    83   25 0.66617148 0.6834855716 0.71568534 0.6348133680  0.66373433
#>    13   84 0.21773799 0.2733261911 0.17634711 0.3023809243  0.20651214
#>    87   33 0.83039311 0.9910996106 0.76090906 0.9753487642  0.63203288
#>    87   69 0.83039311 0.9910996106 0.77090942 0.8868418531  0.80642996
#>    87   51 0.83039311 0.9910996106 0.96650476 0.9310825127  0.87099771
#>    15   79 0.47699423 0.2202285460 0.28601278 0.4093335511  0.36883862
#>    84   93 0.17634711 0.3023809243 0.15552104 0.3316881077  0.16940309
#>     3   16 0.10835562 0.2995831712 0.11077601 0.3887808520  0.11256250
#>   100   11 0.81784737 0.7514760015 0.73987745 0.8719858348  0.78568289
#>    97   73 0.62911235 0.3132129714 0.57361494 0.3256280762  0.61699847
#>    34   12 0.77010333 0.3742238984 0.67124912 0.5799729013  0.72617975
#>    97   75 0.62911235 0.3132129714 0.54321515 0.3670931795  0.63211464
#>    15    2 0.47699423 0.2202285460 0.32761797 0.4311551574  0.41653240
#>   100   24 0.81784737 0.7514760015 0.87020438 0.7138607743  0.82748912
#>    46   64 0.71394613 0.7602742659 0.82917869 0.6667617476  0.75909733
#>    64   77 0.82917869 0.6667617476 0.85370972 0.6674933541  0.84217322
#>    84   86 0.17634711 0.3023809243 0.11962118 0.2103560328  0.14408833
#>    86   99 0.11962118 0.2103560328 0.03587036 0.3664911153 -0.01497774
#>    34   97 0.77010333 0.3742238984 0.62911235 0.3132129714  0.66689957
#>    15   66 0.47699423 0.2202285460 0.38725524 0.4628398578  0.41663670
#>    79    5 0.28601278 0.4093335511 0.26678220 0.1932275582  0.34356200
#>    87   45 0.83039311 0.9910996106 0.93693612 0.9791703724  0.87502438
#>    26   31 0.81701689 0.6319234922 0.92455179 0.5522007592  0.85383197
#>    72   13 0.23259280 0.1440834440 0.21773799 0.2733261911  0.19680565
#>    72   67 0.23259280 0.1440834440 0.15284447 0.1469726250  0.19488134
#>    67   39 0.15284447 0.1469726250 0.04946479 0.0099598093  0.07798922
#>    77   95 0.85370972 0.6674933541 1.00000000 0.6302007247  0.92207289
#>    38   25 0.83900952 0.4979888648 0.71568534 0.6348133680  0.76410076
#>    41   15 0.40908648 0.1343928417 0.47699423 0.2202285460  0.41889557
#>    34   47 0.77010333 0.3742238984 0.59626616 0.5070209645  0.70546712
#>    20   61 0.73691411 0.2315159545 0.62761360 0.2676992779  0.68599298
#>    97   61 0.62911235 0.3132129714 0.62761360 0.2676992779  0.59512895
#>    58   85 0.65112311 0.1975905993 0.47344819 0.1653212179  0.56123883
#>     5   41 0.26678220 0.1932275582 0.40908648 0.1343928417  0.31977753
#>    38   42 0.83900952 0.4979888648 0.93449772 0.4895606211  0.88289771
#>    44   72 0.29827744 0.0196741388 0.23259280 0.1440834440  0.31925030
#>    58   15 0.65112311 0.1975905993 0.47699423 0.2202285460  0.56385230
#>    30   41 0.47963029 0.1401318697 0.40908648 0.1343928417  0.44375489
#>    72   62 0.23259280 0.1440834440 0.17355531 0.0613715131  0.19146814
#>    86    4 0.11962118 0.2103560328 0.00000000 0.0498495959 -0.05620989
#>    41   32 0.40908648 0.1343928417 0.50302728 0.0160756123  0.40089066
#>    63   31 0.98387039 0.5573375645 0.92455179 0.5522007592  0.95003051
#>    20   58 0.73691411 0.2315159545 0.65112311 0.1975905993  0.68130365
#>    34   36 0.77010333 0.3742238984 0.80675176 0.3223175021  0.73140554
#>     4   65 0.00000000 0.0498495959 0.04732251 0.9439205380 -4.77072919
#>    20   92 0.73691411 0.2315159545 0.63749827 0.2940841139  0.71767013
#>    38   10 0.83900952 0.4979888648 0.93920507 0.4176845935  0.85097706
#>    34   96 0.77010333 0.3742238984 0.90269672 0.3325988524  0.84798771
#>    62   68 0.17355531 0.0613715131 0.15852706 0.0000000000  0.10764922
#>     9   30 0.61586053 0.0437430849 0.47963029 0.1401318697  0.57720142
#>    58   88 0.65112311 0.1975905993 0.74369286 0.1621239967  0.69489637
#>    58   35 0.65112311 0.1975905993 0.72829629 0.1360126205  0.69015138
#>    96   21 0.90269672 0.3325988524 0.93136895 0.2961744298  0.89401776
#>    70   20 0.83433588 0.1460736440 0.73691411 0.2315159545  0.81745534
#>    10   29 0.93920507 0.4176845935 0.97361139 0.2803804358  0.97714907
#>    10   63 0.93920507 0.4176845935 0.98387039 0.5573375645  1.20205224
#>    54    9 0.74461723 0.0272072620 0.61586053 0.0437430849  0.68297736
#>    54   81 0.74461723 0.0272072620 0.70681627 0.1208828962  0.74920062
#>     9   58 0.61586053 0.0437430849 0.65112311 0.1975905993  0.62768079
#>    43   29 0.89433514 0.2589198946 0.97361139 0.2803804358  0.93899655
#>    70   88 0.83433588 0.1460736440 0.74369286 0.1621239967  0.79756967
#>    54   14 0.74461723 0.0272072620 0.57273723 0.0001042156  0.66901616
#>    14   48 0.57273723 0.0001042156 0.58000578 0.0332018850  0.60627897
#>    44   14 0.29827744 0.0196741388 0.57273723 0.0001042156  0.36587537
#>    63   95 0.98387039 0.5573375645 1.00000000 0.6302007247  1.94350019
#>    89   50 0.88034794 0.0426704320 0.89797532 0.2317769092  1.02000675
#>    70   91 0.83433588 0.1460736440 0.84401445 0.0047221414  0.80726626
#>    54   35 0.74461723 0.0272072620 0.72829629 0.1360126205  0.78624649
#>    89   70 0.88034794 0.0426704320 0.83433588 0.1460736440  0.94247464
#>     1   49 0.08025720 0.9168790837 0.23826185 0.9713806455  0.13435178
#>    71   90 0.16304251 0.7644512574 0.26054264 0.9287583050  0.21150800
#>    23   65 0.40405074 1.0000000000 0.04732251 0.9439205380 -0.77176007
#>    27   55 0.55645215 0.6620035135 0.47751943 0.8626626106  0.47245781
#>    19   53 0.29883329 0.5941228671 0.35765478 0.5258159409  0.39525008
#>    17   78 0.06122162 0.6125345127 0.05210580 0.5189323421  0.12301227
#>    98   37 0.53198354 0.9318032775 0.69401073 0.9186485788  0.62461630
#>    11   37 0.73987745 0.8719858348 0.69401073 0.9186485788  0.71572458
#>    13   79 0.21773799 0.2733261911 0.28601278 0.4093335511  0.26362445
#>    12   83 0.67124912 0.5799729013 0.66617148 0.6834855716  0.62052478
#>    87   23 0.83039311 0.9910996106 0.40405074 1.0000000000  0.63203288
#>    93    3 0.15552104 0.3316881077 0.10835562 0.2995831712  0.14237958
#>   100   69 0.81784737 0.7514760015 0.77090942 0.8868418531  0.78568289
#>    34   38 0.77010333 0.3742238984 0.83900952 0.4979888648  0.72617975
#>    26   64 0.81701689 0.6319234922 0.82917869 0.6667617476  0.76736158
#>    15   75 0.47699423 0.2202285460 0.54321515 0.3670931795  0.41663670
#>    26   38 0.81701689 0.6319234922 0.83900952 0.4979888648  0.85383197
#>    72    5 0.23259280 0.1440834440 0.26678220 0.1932275582  0.19680565
#>    67    4 0.15284447 0.1469726250 0.00000000 0.0498495959  0.07798922
#>    44   41 0.29827744 0.0196741388 0.40908648 0.1343928417  0.31925030
#>    95   51 1.00000000 0.6302007247 0.96650476 0.9310825127  1.00679698
#>    34   20 0.77010333 0.3742238984 0.73691411 0.2315159545  0.73140554
#>    68   44 0.15852706 0.0000000000 0.29827744 0.0196741388  0.23073448
#>    43   96 0.89433514 0.2589198946 0.90269672 0.3325988524  0.85762375
#>    50   43 0.89797532 0.2317769092 0.89433514 0.2589198946  0.81743962
#>    89   29 0.88034794 0.0426704320 0.97361139 0.2803804358  1.02000675
#>    70   54 0.83433588 0.1460736440 0.74461723 0.0272072620  0.80726626
#>           my1         mx2           my2 bp1 bp2
#>   1.016339535 -0.77176007   7.316839237   0   0
#>   0.911841212  0.21150800   0.846773647   0   0
#>   0.920964370 -0.03235053   0.813307232   0   0
#>   0.813307232  0.13435178   1.016339535   0   0
#>   0.745816386  0.14857297   0.664068678   0   0
#>   0.964948156  0.41867961   0.931843798   0   0
#>   0.832953873  0.35332454   0.815967494   0   0
#>   0.987422534 -0.77176007   7.316839237   0   0
#>   0.822132106  0.24352786   0.740868102   0   0
#>   0.931843798  0.61934378   1.521326468   0   0
#>   0.813326852  0.35332454   0.815967494   0   0
#>   0.690081797  0.24352786   0.740868102   0   0
#>   0.917570289  0.41867961   0.931843798   0   0
#>   0.740868102  0.15174601   0.664782095   0   0
#>   0.744069136  0.33753775   0.722923994   0   0
#>   0.722923994  0.35332454   0.815967494   0   0
#>   0.919857673  0.42990418   0.917570289   0   0
#>   0.886083883  0.61566340   0.958066397   0   0
#>   0.664068678  0.12500976   0.561660795   0   0
#>   0.772254292 -1.46915545   0.714331389   0   0
#>   0.606022415  0.15174601   0.664782095   0   0
#>   0.744817211  0.40444162   0.813326852   0   0
#>   0.617670661  0.40537944   0.509803271   0   0
#>   0.658026773  0.30260897   0.690081797   0   0
#>   0.561660795  0.26097929   0.527043378   0   0
#>   0.559271809  0.11560956   0.469259281   0   0
#>   0.527043378  0.27252437   0.511987250   0   0
#>   0.618449501  0.44227790   0.642247722   0   0
#>   0.774357854  0.59180463   0.772377271   0   0
#>   0.668101583  0.47216653   0.744069136   0   0
#>   1.107869504  0.61934378   1.521326468   0   0
#>   1.068339796  0.74312121   0.928520498   0   0
#>   0.909848128  0.61566340   0.958066397   0   0
#>   0.801551030  0.70647569   0.820873831   0   0
#>   0.820873831  0.71572458   0.894118502   0   0
#>   0.642247722  0.45317681   0.668101583   0   0
#>   0.509803271  0.45279777   0.468709664   0   0
#>   0.468709664  0.47197930   0.425951869   0   0
#>   0.500346189  0.26919398   0.491973503   0   0
#>   0.477110457  0.15744750   0.442704398   0   0
#>   0.469259281  0.05428487   0.441615090   0   0
#>   0.479221775  0.41653240   0.335766789   0   0
#>   0.561154940  0.49883495   0.583410195   0   0
#>   0.446805725  0.55209715   0.561154940   0   0
#>   0.894118502  0.62820614   0.909848128   0   0
#>   0.772377271  0.61925275   0.632186350   0   0
#>   0.335431921  0.25258182   0.348300277   0   0
#>   0.777081344  0.72886152   0.697738528   0   0
#>   1.521326468  0.63203288   1.705017486   0   0
#>   0.928520498  0.80642996   0.935673730   0   0
#>   0.376031787  0.16818829   0.387696023   0   0
#>   0.442704398  0.16818829   0.387696023   0   0
#>   0.632186350  0.45317681   0.668101583   0   0
#>   0.915928217  0.78568289   0.816143772   0   0
#>   0.806093462  0.70647569   0.820873831   0   0
#>   0.344901141  0.05428487   0.441615090   0   0
#>   0.703773285 -1.46915545   0.714331389   0   0
#>   0.629365578  0.61925275   0.632186350   0   0
#>   0.578881850  0.55209715   0.561154940   0   0
#>   0.348300277  0.25449218   0.477110457   0   0
#>   0.631485149  0.76158770   0.552196380   0   0
#>   0.697738528  0.75909733   0.698157673   0   0
#>   0.301343774  0.16940309   0.319499646   0   0
#>   0.700788207  0.82748912   0.709650714   0   0
#>   0.898834951  0.88213641   0.849588916   0   0
#>   0.301990500  0.26919398   0.491973503   0   0
#>   0.319499646  0.18219688   0.376031787   0   0
#>   0.300296378  0.14408833   0.258769933   0   0
#>   0.344100696  0.16818829   0.387696023   0   0
#>   0.389310572  0.49550937   0.300241859   0   0
#>   0.293778126  0.36804593   0.236641505   0   0
#>   0.244770174  0.19085175   0.207293596   0   0
#>   0.479742623  0.76158770   0.552196380   0   0
#>   0.413409142  0.54402753   0.446805725   0   0
#>   0.709650714  0.84051438   0.698305044   0   0
#>   0.698157673  0.76736158   0.668799765   0   0
#>   0.849588916  0.91091379   0.825770350   0   0
#>   0.698305044  0.93123569   0.666032002   0   0
#>   0.642683513  0.87965121   0.604022307   0   0
#>   0.825770350  1.00679698   0.783262688   0   0
#>   0.258769933 -0.01497774   0.238686710   0   0
#>   0.238686710  0.08306412   0.344901141   0   0
#>   0.419304422  0.63211464   0.413409142   0   0
#>   0.335805366  0.47197930   0.425951869   0   0
#>   0.295303799  0.19680565   0.205445227   0   0
#>   0.907966961 15.96497602  10.186596690   0   1
#>   0.569195729  0.88857219   0.514378289   0   0
#>   0.604022307  0.92207289   0.630088508   0   0
#>   0.249889914  0.49550937   0.300241859   0   0
#>   0.205224052  0.19085175   0.207293596   0   0
#>   0.095945108 -9.80070815 -12.153998993   0   1
#>   0.630088508  0.95003051   0.603045525   0   0
#>   0.554461486  0.76736158   0.668799765   0   0
#>   0.196412511  0.56123883   0.187219682   0   0
#>   0.469791028  0.59933118   0.578881850   0   0
#>   0.291226455  0.68599298   0.260872341   0   0
#>   0.291550505  0.55043622   0.249889914   0   0
#>   0.334469546  0.60496969   0.291226455   0   0
#>   0.119894096  0.31925030   0.110291705   0   0
#>   0.666032002  1.00679698   0.783262688   0   0
#>   0.450089039  1.04528998   0.460724535   0   0
#>   0.174098800  0.44375489   0.144680442   0   0
#>   0.207322176  0.55043622   0.249889914   0   0
#>   0.144680442  0.41889557   0.196412511   0   0
#>   0.525206832  0.88857219   0.514378289   0   0
#>   0.106825234  0.17514428   0.028456651   0   0
#>   0.111011459  0.09042251   0.086563876   0   0
#>   0.216569793 -4.77072919   0.750648564   0   0
#>   0.404534303  0.96836797   0.354796138   0   0
#>   0.031433683  0.38906456  -0.641450880   0   0
#>   0.246707079  0.56385230   0.207322176   0   0
#>   0.308010407  0.81602308   0.242929029   0   0
#>   0.086563876  0.10764922   0.044984389   0   0
#>   0.135095276  0.19085175   0.207293596   0   0
#>   0.750648564 -2.40722712   0.703773285   0   0
#>   0.311204816  0.70361011   0.334469546   0   0
#>   0.460724535  0.95677110   0.525206832   0   0
#>   0.410261677  0.85244905   0.404534303   0   0
#>   0.390323057  0.85762375   0.300400092   0   0
#>   0.044984389 -0.72276546  -9.048253519   0   1
#>   0.028456651  0.23073448  -0.006729326   0   0
#>   0.133568869  0.56362185   0.174098800   0   0
#>   0.070272972  0.52807952   0.085034923   0   0
#>   0.079168461  0.19733307   0.106825234   0   0
#>  -0.006729326  0.26030165   0.079168461   0   0
#>   0.242929029  0.85762375   0.300400092   0   0
#>   0.193126232  0.79756967   0.202414184   0   0
#>   0.234791749  0.81602308   0.242929029   0   0
#>   0.173301851  0.78319535   0.121236389   0   0
#>   0.167355144  0.74920062   0.083521536   0   0
#>  -0.641450880  0.54159401   0.024290508   0   0
#>   0.354796138  0.97714907   0.354229850   0   0
#>   0.296269856  0.93882483   0.251728263   0   0
#>   0.225087927  0.81743962   0.234791749   0   0
#>   0.354229850  0.93882483   0.251728263   0   0
#>   0.410587141  1.94350019   0.383122557   0   0
#>   0.126353639  0.52807952   0.085034923   0   0
#>   0.056798439  0.62768079   0.121998758   0   0
#>   0.121998758  0.69015138   0.167355144   0   0
#>   0.251093907  1.02000675   0.125027067   0   0
#>  -0.051910887  0.60627897   0.010085096   0   0
#>   0.010085096  0.57209633   0.126353639   0   0
#>   0.024290508  0.52807952   0.085034923   0   0
#>  -0.966669251  0.36648280  -3.380902454   0   1
#>   0.383122557  4.78593622   0.168704066   0   1
#>   0.121236389  0.78624649   0.089078461   0   0
#>   0.073213037  0.71417578  -0.338300673   0   0
#>   0.132254149  0.81745534   0.225087927   0   0
#>   0.073417360 14.55866001 -13.089959288   0   1
#>  -0.338300673  0.77690855  -4.023498745   0   1
#>   0.911841212  0.32085727   0.987422534   0   0
#>   0.920964370  0.17627298   0.911841212   0   0
#>   0.813307232 -0.08814326   0.772254292   0   0
#>   0.846773647  0.02660541   0.745816386   0   0
#>   0.964948156  0.32085727   0.987422534   0   0
#>   0.832953873  0.33201422   0.964948156   0   0
#>   0.822132106  0.28279098   0.832953873   0   0
#>   0.917570289  0.40444162   0.813326852   0   0
#>   0.722923994  0.30260897   0.690081797   0   0
#>   0.919857673  0.51890491   0.886083883   0   0
#>   0.664068678  0.15174601   0.664782095   0   0
#>   0.772254292  0.02660541   0.745816386   0   0
#>   0.606022415  0.26097929   0.527043378   0   0
#>   0.744817211  0.47216653   0.744069136   0   0
#>   0.617670661  0.39555913   0.618449501   0   0
#>   0.658026773  0.24875505   0.606022415   0   0
#>   0.559271809  0.12500976   0.561660795   0   0
#>   0.618449501  0.31869952   0.658026773   0   0
#>   0.511987250  0.27374272   0.500346189   0   0
#>   0.774357854  0.51890491   0.886083883   0   0
#>   1.107869504  0.47603003   0.919857673   0   0
#>   1.068339796  0.61709708   1.107869504   0   0
#>   0.909848128  0.62210327   0.801551030   0   0
#>   0.468709664  0.49883495   0.583410195   0   0
#>   0.500346189  0.34031616   0.479221775   0   0
#>   0.469259281  0.15744750   0.442704398   0   0
#>   0.479221775  0.40537944   0.509803271   0   0
#>   0.583410195  0.44227790   0.642247722   0   0
#>   0.446805725  0.47197930   0.425951869   0   0
#>   0.772377271  0.60133297   0.777081344   0   0
#>   0.335431921  0.20651214   0.301343774   0   0
#>   0.491973503  0.25449218   0.477110457   0   0
#>   0.777081344  0.62210327   0.801551030   0   0
#>   0.915928217  0.74312121   0.928520498   0   0
#>   0.806093462  0.76123204   0.700788207   0   0
#>   0.441615090 -2.40722712   0.703773285   0   0
#>   0.629365578  0.59933118   0.578881850   0   0
#>   0.348300277  0.18219688   0.376031787   0   0
#>   0.631485149  0.72886152   0.697738528   0   0
#>   0.301343774  0.16679975   0.244770174   0   0
#>   1.705017486  0.80642996   0.935673730   0   0
#>   0.935673730  0.87099771   0.898834951   0   0
#>   0.898834951  0.87502438   0.907966961   0   0
#>   0.301990500  0.36070694   0.293778126   0   0
#>   0.319499646  0.14237958   0.300296378   0   0
#>   0.344100696  0.08306412   0.344901141   0   0
#>   0.816143772  0.77014919   0.806093462   0   0
#>   0.389310572  0.59512895   0.291550505   0   0
#>   0.479742623  0.70546712   0.469791028   0   0
#>   0.413409142  0.61699847   0.389310572   0   0
#>   0.335766789  0.36883862   0.301990500   0   0
#>   0.709650714  0.91091379   0.825770350   0   0
#>   0.698157673  0.76123204   0.700788207   0   0
#>   0.642683513  0.84051438   0.698305044   0   0
#>   0.258769933  0.16679975   0.244770174   0   0
#>   0.238686710 -0.05620989   0.216569793   0   0
#>   0.419304422  0.70361011   0.334469546   0   0
#>   0.335805366  0.41653240   0.335766789   0   0
#>   0.295303799  0.36070694   0.293778126   0   0
#>   0.907966961  1.91291626  10.177638875   0   1
#>   0.569195729  0.87965121   0.604022307   0   0
#>   0.205445227  0.19488134   0.205224052   0   0
#>   0.205224052  0.19146814   0.111011459   0   0
#>   0.095945108  0.09042251   0.086563876   0   0
#>   0.630088508  0.93123569   0.666032002   0   0
#>   0.554461486  0.76158770   0.552196380   0   0
#>   0.196412511  0.36804593   0.236641505   0   0
#>   0.469791028  0.66689957   0.419304422   0   0
#>   0.260872341  0.68130365   0.246707079   0   0
#>   0.291550505  0.60496969   0.291226455   0   0
#>   0.187219682  0.56362185   0.174098800   0   0
#>   0.119894096  0.36804593   0.236641505   0   0
#>   0.450089039  0.88857219   0.514378289   0   0
#>   0.110291705  0.26030165   0.079168461   0   0
#>   0.207322176  0.56123883   0.187219682   0   0
#>   0.144680442  0.44980824   0.070272972   0   0
#>   0.111011459  0.19733307   0.106825234   0   0
#>   0.216569793  0.05311176   0.135095276   0   0
#>   0.031433683  0.44980824   0.070272972   0   0
#>   0.603045525  0.95677110   0.525206832   0   0
#>   0.246707079  0.70249176   0.193126232   0   0
#>   0.308010407  0.84798771   0.390323057   0   0
#>   0.750648564 -5.88608399   0.809683459   0   1
#>   0.311204816  0.68599298   0.260872341   0   0
#>   0.410261677  0.88289771   0.450089039   0   0
#>   0.390323057  0.85244905   0.404534303   0   0
#>   0.044984389  0.17514428   0.028456651   0   0
#>   0.133568869  0.57209633   0.126353639   0   0
#>   0.173301851  0.70249176   0.193126232   0   0
#>   0.167355144  0.69489637   0.173301851   0   0
#>   0.296269856  0.96836797   0.354796138   0   0
#>   0.225087927  0.79756967   0.202414184   0   0
#>   0.354229850  1.20205224   0.410587141   0   0
#>   0.410587141  1.04528998   0.460724535   0   0
#>   0.056798439  0.66901616  -0.051910887   0   0
#>   0.083521536  0.68297736   0.056798439   0   0
#>   0.121998758  0.57720142   0.133568869   0   0
#>   0.251093907  0.93882483   0.251728263   0   0
#>   0.202414184  0.78319535   0.121236389   0   0
#>  -0.051910887  0.71417578  -0.338300673   0   0
#>   0.010085096  0.54159401   0.024290508   0   0
#>  -0.966669251  0.38906456  -0.641450880   0   0
#>   0.383122557  0.95003051   0.603045525   0   0
#>   0.125027067  0.94247464   0.132254149   0   0
#>   0.073213037  0.81025032   0.073417360   0   0
#>   0.089078461  0.74920062   0.083521536   0   0
#>   0.132254149  0.81025032   0.073417360   0   0
#>   1.016339535  0.16725015   0.920964370   0   0
#>   0.846773647  0.25303388   0.822132106   0   0
#>   7.316839237 -1.20181749  10.052486467   0   1
#>   0.744817211  0.54755465   0.774357854   0   0
#>   0.617670661  0.27252437   0.511987250   0   0
#>   0.559271809 -1.46915545   0.714331389   0   0
#>   1.068339796  0.61566340   0.958066397   0   0
#>   0.894118502  0.73791281   0.915928217   0   0
#>   0.335431921  0.34356200   0.295303799   0   0
#>   0.629365578  0.66373433   0.631485149   0   0
#>   1.705017486  0.68097726   4.049528513   0   1
#>   0.300296378  0.11256250   0.344100696   0   0
#>   0.816143772  0.88213641   0.849588916   0   0
#>   0.479742623  0.85097706   0.410261677   0   0
#>   0.668799765  0.84217322   0.642683513   0   0
#>   0.335805366  0.49550937   0.300241859   0   0
#>   0.569195729  0.76410076   0.554461486   0   0
#>   0.205445227  0.31977753   0.119894096   0   0
#>   0.095945108  0.05311176   0.135095276   0   0
#>   0.110291705  0.40089066   0.031433683   0   0
#>   0.783262688  4.28967686   1.148724622   0   1
#>   0.308010407  0.71767013   0.311204816   0   0
#>  -0.006729326  0.36587537  -0.966669251   0   0
#>   0.300400092  0.89401776   0.296269856   0   0
#>   0.234791749  0.93899655   0.251093907   0   0
#>   0.125027067  4.66563551  -1.305303622   0   1
#>   0.073213037  0.78624649   0.089078461   0   0
#> 
#> $x
#>              [,1]         [,2]
#>   [1,] 0.08025720 0.9168790837
#>   [2,] 0.32761797 0.4311551574
#>   [3,] 0.10835562 0.2995831712
#>   [4,] 0.00000000 0.0498495959
#>   [5,] 0.26678220 0.1932275582
#>   [6,] 0.56645419 0.8854680274
#>   [7,] 0.22097551 0.4102981086
#>   [8,] 0.20461746 0.5795403111
#>   [9,] 0.61586053 0.0437430849
#>  [10,] 0.93920507 0.4176845935
#>  [11,] 0.73987745 0.8719858348
#>  [12,] 0.67124912 0.5799729013
#>  [13,] 0.21773799 0.2733261911
#>  [14,] 0.57273723 0.0001042156
#>  [15,] 0.47699423 0.2202285460
#>  [16,] 0.11077601 0.3887808520
#>  [17,] 0.06122162 0.6125345127
#>  [18,] 0.47185342 0.9769194759
#>  [19,] 0.29883329 0.5941228671
#>  [20,] 0.73691411 0.2315159545
#>  [21,] 0.93136895 0.2961744298
#>  [22,] 0.32728318 0.7452388683
#>  [23,] 0.40405074 1.0000000000
#>  [24,] 0.87020438 0.7138607743
#>  [25,] 0.71568534 0.6348133680
#>  [26,] 0.81701689 0.6319234922
#>  [27,] 0.55645215 0.6620035135
#>  [28,] 0.68682803 0.8786845954
#>  [29,] 0.97361139 0.2803804358
#>  [30,] 0.47963029 0.1401318697
#>  [31,] 0.92455179 0.5522007592
#>  [32,] 0.50302728 0.0160756123
#>  [33,] 0.76090906 0.9753487642
#>  [34,] 0.77010333 0.3742238984
#>  [35,] 0.72829629 0.1360126205
#>  [36,] 0.80675176 0.3223175021
#>  [37,] 0.69401073 0.9186485788
#>  [38,] 0.83900952 0.4979888648
#>  [39,] 0.04946479 0.0099598093
#>  [40,] 0.50076329 0.5137606018
#>  [41,] 0.40908648 0.1343928417
#>  [42,] 0.93449772 0.4895606211
#>  [43,] 0.89433514 0.2589198946
#>  [44,] 0.29827744 0.0196741388
#>  [45,] 0.93693612 0.9791703724
#>  [46,] 0.71394613 0.7602742659
#>  [47,] 0.59626616 0.5070209645
#>  [48,] 0.58000578 0.0332018850
#>  [49,] 0.23826185 0.9713806455
#>  [50,] 0.89797532 0.2317769092
#>  [51,] 0.96650476 0.9310825127
#>  [52,] 0.25178173 0.6574059527
#>  [53,] 0.35765478 0.5258159409
#>  [54,] 0.74461723 0.0272072620
#>  [55,] 0.47751943 0.8626626106
#>  [56,] 0.35457890 0.7406075735
#>  [57,] 0.44928869 0.5344208849
#>  [58,] 0.65112311 0.1975905993
#>  [59,] 0.36233749 0.8907970183
#>  [60,] 0.19375275 0.0556529365
#>  [61,] 0.62761360 0.2676992779
#>  [62,] 0.17355531 0.0613715131
#>  [63,] 0.98387039 0.5573375645
#>  [64,] 0.82917869 0.6667617476
#>  [65,] 0.04732251 0.9439205380
#>  [66,] 0.38725524 0.4628398578
#>  [67,] 0.15284447 0.1469726250
#>  [68,] 0.15852706 0.0000000000
#>  [69,] 0.77090942 0.8868418531
#>  [70,] 0.83433588 0.1460736440
#>  [71,] 0.16304251 0.7644512574
#>  [72,] 0.23259280 0.1440834440
#>  [73,] 0.57361494 0.3256280762
#>  [74,] 0.18718042 0.3245232712
#>  [75,] 0.54321515 0.3670931795
#>  [76,] 0.11605180 0.8505144414
#>  [77,] 0.85370972 0.6674933541
#>  [78,] 0.05210580 0.5189323421
#>  [79,] 0.28601278 0.4093335511
#>  [80,] 0.54974355 0.6309822231
#>  [81,] 0.70681627 0.1208828962
#>  [82,] 0.18637549 0.5078898838
#>  [83,] 0.66617148 0.6834855716
#>  [84,] 0.17634711 0.3023809243
#>  [85,] 0.47344819 0.1653212179
#>  [86,] 0.11962118 0.2103560328
#>  [87,] 0.83039311 0.9910996106
#>  [88,] 0.74369286 0.1621239967
#>  [89,] 0.88034794 0.0426704320
#>  [90,] 0.26054264 0.9287583050
#>  [91,] 0.84401445 0.0047221414
#>  [92,] 0.63749827 0.2940841139
#>  [93,] 0.15552104 0.3316881077
#>  [94,] 0.35917950 0.7113160521
#>  [95,] 1.00000000 0.6302007247
#>  [96,] 0.90269672 0.3325988524
#>  [97,] 0.62911235 0.3132129714
#>  [98,] 0.53198354 0.9318032775
#>  [99,] 0.03587036 0.3664911153
#> [100,] 0.81784737 0.7514760015
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,49,65), [9,4,3]
#> 2: (49,76,90), [5,10,3]
#> 3: (49,1,76), [4,2,1]
#> 4: (76,1,65), [1,23,3]
#> 5: (71,90,76), [2,6,12]
#> 6: (76,17,71), [22,5,23]
#> 7: (90,59,23), [13,10,8]
#> 8: (90,22,59), [11,7,12]
#> 9: (23,65,49), [1,10,0]
#> 10: (90,23,49), [9,2,7]
#> 11: (56,59,22), [8,19,14]
#> 12: (90,71,22), [17,8,5]
#> 13: (59,18,23), [58,7,16]
#> 14: (55,59,56), [11,26,16]
#> 15: (94,22,52), [17,28,19]
#> 16: (55,18,59), [13,14,20]
#> 17: (22,71,52), [25,15,12]
#> 18: (27,56,94), [19,36,26]
#> 19: (94,56,22), [11,15,18]
#> 20: (98,18,55), [16,21,37]
#> 21: (55,6,98), [39,20,35]
#> 22: (71,17,8), [29,25,6]
#> 23: (76,65,17), [31,6,4]
#> 24: (19,52,8), [25,32,28]
#> 25: (71,8,52), [24,17,22]
#> 26: (27,55,56), [14,18,35]
#> 27: (19,53,57), [44,33,34]
#> 28: (19,94,52), [15,24,33]
#> 29: (17,82,8), [32,22,30]
#> 30: (17,78,82), [48,29,31]
#> 31: (78,17,65), [23,67,30]
#> 32: (8,82,19), [34,24,29]
#> 33: (19,57,94), [43,28,27]
#> 34: (53,19,82), [32,46,27]
#> 35: (55,27,6), [54,21,26]
#> 36: (80,27,94), [18,43,62]
#> 37: (98,33,18), [58,20,38]
#> 38: (98,37,33), [59,37,39]
#> 39: (37,98,6), [21,40,38]
#> 40: (28,37,6), [39,41,53]
#> 41: (6,46,28), [42,40,57]
#> 42: (46,11,28), [53,41,64]
#> 43: (57,80,94), [36,33,51]
#> 44: (53,66,57), [45,27,49]
#> 45: (57,66,40), [68,51,44]
#> 46: (53,82,2), [56,49,34]
#> 47: (79,82,7), [61,72,56]
#> 48: (82,78,16), [66,61,30]
#> 49: (53,2,66), [90,44,46]
#> 50: (47,80,40), [51,52,71]
#> 51: (57,40,80), [50,43,45]
#> 52: (75,47,40), [50,68,89]
#> 53: (11,37,28), [40,42,63]
#> 54: (6,27,83), [62,57,35]
#> 55: (13,79,74), [72,75,104]
#> 56: (79,2,82), [46,47,80]
#> 57: (6,83,46), [74,41,54]
#> 58: (18,33,23), [77,13,37]
#> 59: (37,69,33), [78,38,63]
#> 60: (74,7,93), [69,81,72]
#> 61: (82,16,7), [69,47,48]
#> 62: (83,27,80), [36,70,54]
#> 63: (37,11,69), [84,59,53]
#> 64: (100,11,46), [42,76,84]
#> 65: (3,16,99), [66,100,83]
#> 66: (99,16,78), [48,67,65]
#> 67: (99,78,65), [31,143,66]
#> 68: (75,40,66), [45,103,52]
#> 69: (93,7,16), [61,83,60]
#> 70: (12,83,80), [62,71,73]
#> 71: (12,80,47), [50,116,70]
#> 72: (74,79,7), [47,60,55]
#> 73: (83,12,25), [102,74,70]
#> 74: (83,25,46), [92,57,73]
#> 75: (13,74,84), [81,87,55]
#> 76: (46,64,100), [91,64,92]
#> 77: (87,23,33), [58,78,0]
#> 78: (87,33,69), [59,79,77]
#> 79: (87,69,51), [94,105,78]
#> 80: (15,2,79), [56,86,90]
#> 81: (84,74,93), [60,82,75]
#> 82: (93,3,84), [99,81,83]
#> 83: (3,93,16), [69,65,82]
#> 84: (100,69,11), [63,64,94]
#> 85: (97,75,73), [109,119,89]
#> 86: (79,5,15), [125,80,104]
#> 87: (84,86,13), [98,75,99]
#> 88: (34,38,12), [102,116,146]
#> 89: (97,47,75), [52,85,101]
#> 90: (15,66,2), [49,80,103]
#> 91: (100,64,24), [95,97,76]
#> 92: (46,25,64), [93,76,74]
#> 93: (26,64,25), [92,114,96]
#> 94: (69,100,51), [97,79,84]
#> 95: (64,77,24), [123,91,96]
#> 96: (64,26,77), [107,95,93]
#> 97: (100,24,51), [130,94,91]
#> 98: (67,13,86), [87,142,111]
#> 99: (84,3,86), [100,87,82]
#> 100: (86,3,99), [65,134,99]
#> 101: (34,47,97), [89,120,116]
#> 102: (38,25,12), [73,88,114]
#> 103: (15,75,66), [68,90,109]
#> 104: (79,13,5), [110,86,55]
#> 105: (87,51,45), [0,0,79]
#> 106: (26,38,31), [135,107,114]
#> 107: (26,31,77), [113,96,106]
#> 108: (61,73,15), [109,128,119]
#> 109: (75,15,73), [108,85,103]
#> 110: (72,5,13), [104,111,122]
#> 111: (72,13,67), [98,133,110]
#> 112: (67,4,39), [0,141,142]
#> 113: (77,31,95), [138,123,107]
#> 114: (38,26,25), [93,102,106]
#> 115: (41,85,15), [121,125,129]
#> 116: (34,12,47), [71,101,88]
#> 117: (97,61,92), [118,120,119]
#> 118: (20,92,61), [117,139,144]
#> 119: (97,73,61), [108,117,85]
#> 120: (34,97,92), [117,144,101]
#> 121: (58,15,85), [115,127,128]
#> 122: (5,72,41), [126,125,110]
#> 123: (77,95,24), [130,95,113]
#> 124: (38,10,42), [145,135,146]
#> 125: (41,15,5), [86,122,115]
#> 126: (44,41,72), [122,152,137]
#> 127: (58,85,30), [129,150,121]
#> 128: (58,61,15), [108,121,139]
#> 129: (30,85,41), [115,151,127]
#> 130: (95,51,24), [97,123,0]
#> 131: (63,31,42), [135,145,138]
#> 132: (72,62,60), [149,152,133]
#> 133: (72,67,62), [141,132,111]
#> 134: (86,99,4), [143,142,100]
#> 135: (42,31,38), [106,124,131]
#> 136: (34,96,10), [162,146,147]
#> 137: (41,44,32), [161,151,126]
#> 138: (63,95,31), [113,131,177]
#> 139: (20,61,58), [128,156,118]
#> 140: (34,20,36), [154,147,144]
#> 141: (67,39,62), [148,133,112]
#> 142: (4,67,86), [98,134,112]
#> 143: (4,99,65), [67,0,134]
#> 144: (20,34,92), [120,118,140]
#> 145: (10,63,42), [131,124,166]
#> 146: (38,34,10), [136,124,88]
#> 147: (34,36,96), [155,136,140]
#> 148: (62,39,68), [0,149,141]
#> 149: (62,68,60), [153,132,148]
#> 150: (9,58,30), [127,167,170]
#> 151: (41,32,30), [160,129,137]
#> 152: (44,72,60), [132,153,126]
#> 153: (68,44,60), [152,149,176]
#> 154: (20,43,36), [155,140,157]
#> 155: (43,96,36), [147,154,163]
#> 156: (58,88,20), [172,139,158]
#> 157: (50,43,20), [154,164,171]
#> 158: (58,35,88), [180,156,159]
#> 159: (58,81,35), [169,158,170]
#> 160: (48,30,32), [151,175,167]
#> 161: (44,14,32), [175,137,176]
#> 162: (96,21,10), [165,136,163]
#> 163: (96,43,21), [178,162,155]
#> 164: (70,50,20), [157,172,183]
#> 165: (10,21,29), [178,166,162]
#> 166: (10,29,63), [177,145,165]
#> 167: (9,30,48), [160,174,150]
#> 168: (54,81,9), [170,173,169]
#> 169: (54,35,81), [159,168,182]
#> 170: (9,81,58), [159,150,168]
#> 171: (43,50,29), [179,178,157]
#> 172: (70,20,88), [156,180,164]
#> 173: (54,9,14), [174,185,168]
#> 174: (14,9,48), [167,175,173]
#> 175: (14,48,32), [160,161,174]
#> 176: (44,68,14), [0,161,153]
#> 177: (63,29,95), [0,138,166]
#> 178: (43,29,21), [165,163,171]
#> 179: (89,29,50), [171,183,0]
#> 180: (88,35,70), [182,172,158]
#> 181: (70,54,91), [185,184,182]
#> 182: (54,70,35), [180,169,181]
#> 183: (89,50,70), [164,184,179]
#> 184: (70,91,89), [0,183,181]
#> 185: (54,14,91), [0,181,173]
#> boundary nodes:  23 65 4 39 68 14 91 89 29 95 51 45 87 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.19290384 0.16431587 0.07540238 0.04261378 0.18288300 0.11689862 0.14971967 
#>                                                                              
#> 0.16824097 0.16536055 0.07162342 0.15038971 0.11582353 0.13932271 0.13904440 
#>                                                                              
#> 0.02965061 0.02768584 0.11439727 0.05775103 0.14714274 0.33167738 0.09103585 
#>                                                                              
#> 0.17323916 0.09203705 0.12016897 0.07393614 0.13472300 0.14171409 0.19852346 
#>                                                                              
#> 0.22368824 0.20334263 0.28905990 0.08769434 0.13180144 0.12147591 0.05347069 
#>                                                                              
#> 0.20680469 0.09472060 0.12440659 0.16074096 0.10354379 0.14276416 0.06753165 
#>                                                                              
#> 0.12704320 0.09574039 0.04060432 0.11180255 0.13023294 0.09043738 0.35770874 
#>                                                                              
#> 0.08907009 0.10229282 0.11228055 0.03173840 0.03440471 0.11468177 0.07815171 
#>                                                                              
#> 0.42501511 0.12771869 0.13240373 0.06504442 0.07058367 0.12547295 0.02465044 
#>                                                                              
#> 0.08546873 0.20053622 0.04698057 0.03245999 0.06804903 0.07253757 0.05141500 
#>                                                                              
#> 0.21193902 0.11658538 0.18672154 0.14964690 0.06246139 0.11790437 0.23314699 
#>                                                                              
#> 0.04921394 0.05110361 0.23761112 0.08993551 0.09864474 0.19657164 0.18300525 
#>                                                                              
#> 0.09392085 0.05645128 0.10127394 0.13531808 0.14298470 0.14204363 0.06354495 
#>                                                                              
#> 0.10851924 0.10137275 0.05502171 0.10461563 0.02817563 0.07919344 0.02088628 
#>                                                                              
#> 0.17675952 0.15442120 0.07203001 0.02593688 0.15792298 0.07140725 0.06342482 
#>                                                                              
#> 0.02099140 0.08807093 0.31866681 0.09258749 0.20478146 0.07394541 0.11455225 
#>                                                                              
#> 0.13431910 0.07156286 0.57754298 0.15494027 0.08385329 0.17459734 0.09649426 
#>                                                                              
#> 0.10951611 0.06586424 0.18086263 0.12624331 0.09658418 0.11054359 0.10812079 
#>                                                                              
#> 0.06972227 0.15978847 0.03031267 0.02627356 0.07151617 0.12176258 0.05253002 
#>                                                                              
#> 0.16106142 0.04509849 0.27714707 0.14666036 0.11926230 0.09479346 0.08990616 
#>                                                                              
#> 0.06135120 0.03737218 0.07886063 0.41421018 0.35081419 0.10651581 0.10190873 
#>                                                                              
#> 0.10674750 0.05253756 0.27131653 0.04809467 0.17188349 0.11596730 0.09805610 
#>                                                                              
#> 0.16021846 0.10864277 0.19527847 0.11856822 0.04656322 0.08801575 0.18952714 
#>                                                                              
#> 0.24421465 0.09533767 0.21663658 0.16186770 0.07885809 0.16313795 0.13181771 
#>                                                                              
#> 0.17221480 0.09181219 0.07517382 0.23303025 0.12056482 0.05546605 0.09931200 
#>                                                                              
#> 0.14107531 0.06958573 0.13933870 0.15268757 0.22525641 0.05962304 0.14014614 
#>                                                                              
#> 0.19346149 0.08321704 0.10427309 0.15330335 0.13177843 0.09219236 0.06943058 
#>                                                                              
#> 0.05057057 0.07124691 0.12003328 0.14875629 0.26876498 0.03595325 0.08923051 
#>                                                                              
#> 0.14353372 0.05686912 0.22826477 0.10139726 0.25846335 0.06446831 0.14840194 
#>                                                                              
#> 0.02454193 0.10810371 0.17717890 0.15362549 0.25867612 0.21695994 0.10720877 
#>                                                                              
#> 0.13386363 0.13009363 0.07980065 0.17163878 0.15096883 0.18420043 0.10944967 
#>                                                                              
#> 0.21875654 0.11513398 0.04553836 0.18058153 0.15398711 0.09585944 0.14068457 
#>                                                                              
#> 0.17559426 0.07077688 0.10162032 0.20017878 0.15107562 0.05954059 0.09225522 
#>                                                                              
#> 0.06354039 0.89532244 0.11746609 0.12840531 0.13897356 0.06318474 0.16688162 
#>                                                                              
#> 0.09913142 0.09872967 0.04635553 0.12958160 0.14154938 0.14662177 0.12981418 
#>                                                                              
#> 0.10101503 0.15783696 0.08212965 0.09205309 0.17400376 0.03388639 0.27515660 
#>                                                                              
#> 0.07462710 0.18992626 0.14168247 0.11002263 0.11317833 0.16714033 0.19105780 
#>                                                                              
#> 0.36110932 0.21562571 0.09014324 0.09404501 0.16256031 0.06543063 0.15218229 
#>                                                                              
#> 0.10363713 0.42643526 0.05705527 0.14327276 0.14165391 0.03690005 0.16110379 
#>                                                                              
#> 0.13572826 0.05986701 0.18109200 0.15949615 0.30274045 0.14651649 0.14112845 
#>                                             
#> 0.07415190 0.02738599 0.25535097 0.14892499 
#> 
#> $alpha
#> [1] 0.2878548
#> 
#> attr(,"class")
#> [1] "scree"