Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, global_rng = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

global_rng

an optional vector c(min, max) that specifies the global range of the data set if you want all variables one the same scale

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1         y1          x2         y2          mx1
#>    88   85 0.038138308 0.28130555 0.000000000 0.23317149  0.023979680
#>    85   52 0.000000000 0.23317149 0.009474413 0.17004206 -0.001752625
#>    52   96 0.009474413 0.17004206 0.036748940 0.15072695  0.043403487
#>    96   15 0.036748940 0.15072695 0.129978413 0.21895214  0.075678350
#>    15   88 0.129978413 0.21895214 0.038138308 0.28130555  0.076951577
#>    70  100 0.836667421 0.37329465 0.809837702 0.39360708  0.824426024
#>   100   17 0.809837702 0.39360708 0.701320441 0.33106643  0.747620619
#>    17   42 0.701320441 0.33106643 0.809663713 0.33765366  0.753963825
#>    42   70 0.809663713 0.33765366 0.836667421 0.37329465  0.824418505
#>    30   75 0.641509881 0.44241556 0.668376406 0.50457353  0.516896719
#>    75   95 0.668376406 0.50457353 0.627437980 0.66958687  0.536497231
#>    95   49 0.627437980 0.66958687 0.472123903 0.72748893  0.536855260
#>    49   89 0.472123903 0.72748893 0.440003910 0.74627100  0.439120541
#>    89   64 0.440003910 0.74627100 0.403310175 0.74004239  0.427800379
#>    64    5 0.403310175 0.74004239 0.288583668 0.68598089  0.366565037
#>     5    4 0.288583668 0.68598089 0.300162730 0.59947505  0.364855566
#>     4   14 0.300162730 0.59947505 0.281220730 0.53113595  0.424894273
#>    14    9 0.281220730 0.53113595 0.314380451 0.37780051  0.457782908
#>     9   30 0.314380451 0.37780051 0.641509881 0.44241556  0.462796194
#>     5   94 0.288583668 0.68598089 0.114516864 0.57436537  0.201473369
#>    94   97 0.114516864 0.57436537 0.271888992 0.53096363  0.205796716
#>    97   14 0.271888992 0.53096363 0.281220730 0.53113595  0.275846292
#>    64   92 0.403310175 0.74004239 0.342701871 0.83216746  0.326058237
#>    92   34 0.342701871 0.83216746 0.199903247 0.78310528  0.283957227
#>    34   38 0.199903247 0.78310528 0.128040002 0.70959266  0.208982648
#>    38   94 0.128040002 0.70959266 0.114516864 0.57436537  0.198961427
#>    80   74 0.240718734 0.96686189 0.158197512 0.99102621  0.190464710
#>    74   22 0.158197512 0.99102621 0.034555363 0.92425262  0.081372185
#>    22   44 0.034555363 0.92425262 0.110211316 0.85610656  0.088635348
#>    44   79 0.110211316 0.85610656 0.245192825 0.93081229  0.190111975
#>    79   80 0.245192825 0.93081229 0.240718734 0.96686189  0.192293158
#>    77   71 0.197849803 0.18381704 0.255938815 0.09020146  0.274315866
#>    71   87 0.255938815 0.09020146 0.254336020 0.08170851  0.304796729
#>    87   93 0.254336020 0.08170851 0.381749537 0.07421042  0.316400710
#>    93   12 0.381749537 0.07421042 0.429521324 0.11991507  0.382315943
#>    12   69 0.429521324 0.11991507 0.360440337 0.25153041  0.380631539
#>    69   77 0.360440337 0.25153041 0.197849803 0.18381704  0.288399097
#>    98   94 0.150207681 0.45489951 0.114516864 0.57436537  0.128420008
#>    94   67 0.114516864 0.57436537 0.060594719 0.53510635  0.117444593
#>    67   46 0.060594719 0.53510635 0.010056833 0.45758950  0.062041865
#>    46   59 0.010056833 0.45758950 0.001605756 0.42266691  0.059796815
#>    59   91 0.001605756 0.42266691 0.041737815 0.35170314  0.064176708
#>    91   25 0.041737815 0.35170314 0.182150030 0.33300439  0.116308957
#>    25   98 0.182150030 0.33300439 0.150207681 0.45489951  0.134183974
#>    97   98 0.271888992 0.53096363 0.150207681 0.45489951  0.243716225
#>    98   66 0.150207681 0.45489951 0.215557434 0.33274457  0.234743833
#>    66   69 0.215557434 0.33274457 0.360440337 0.25153041  0.291172533
#>    69   24 0.360440337 0.25153041 0.448709294 0.22997522  0.430957076
#>    24   30 0.448709294 0.22997522 0.641509881 0.44241556  0.481063545
#>    72   90 0.843603004 0.12378220 0.956687703 0.17536033  0.880379755
#>    90   40 0.956687703 0.17536033 1.000000000 0.38625161  1.230005327
#>    40   48 1.000000000 0.38625161 0.982352296 0.36026156  1.182136379
#>    48   29 0.982352296 0.36026156 0.892295901 0.37562867  0.926131326
#>    29   19 0.892295901 0.37562867 0.824561414 0.32423770  0.901812693
#>    19   56 0.824561414 0.32423770 0.758204331 0.20620359  0.853383424
#>    56   72 0.758204331 0.20620359 0.843603004 0.12378220  0.854388745
#>    76   23 0.025429672 0.77741184 0.025348304 0.75345982  0.021335658
#>    23   83 0.025348304 0.75345982 0.022419730 0.69681308  0.039766802
#>    83   45 0.022419730 0.69681308 0.103152743 0.70658208  0.060272878
#>    45   76 0.103152743 0.70658208 0.025429672 0.77741184  0.091773738
#>    93   60 0.381749537 0.07421042 0.467487930 0.09772376  0.427981189
#>    60   55 0.467487930 0.09772376 0.529143180 0.22063691  0.482559817
#>    55   24 0.529143180 0.22063691 0.448709294 0.22997522  0.482207156
#>    57   50 0.490619193 0.00000000 0.886639057 0.01361826  0.693372848
#>    50   65 0.886639057 0.01361826 0.789142238 0.10541048  0.721292455
#>    65   56 0.789142238 0.10541048 0.758204331 0.20620359  0.725574156
#>    56   82 0.758204331 0.20620359 0.673591737 0.21986954  0.707220025
#>    82   78 0.673591737 0.21986954 0.614844107 0.12126683  0.652406836
#>    78   57 0.614844107 0.12126683 0.490619193 0.00000000  0.606322286
#>    96   77 0.036748940 0.15072695 0.197849803 0.18381704  0.125524670
#>    77   51 0.197849803 0.18381704 0.200660136 0.27178338  0.177941223
#>    51   25 0.200660136 0.27178338 0.182150030 0.33300439  0.135360536
#>    25   88 0.182150030 0.33300439 0.038138308 0.28130555  0.114512961
#>    17   82 0.701320441 0.33106643 0.673591737 0.21986954  0.690136085
#>    19   42 0.824561414 0.32423770 0.809663713 0.33765366  0.788019521
#>   100   43 0.809837702 0.39360708 0.794290738 0.42310589  0.746065997
#>    43   86 0.794290738 0.42310589 0.673476387 0.39050246  0.739691550
#>    86   99 0.673476387 0.39050246 0.654417496 0.40436483  0.611391965
#>    99   55 0.654417496 0.40436483 0.529143180 0.22063691  0.591683897
#>    55   82 0.529143180 0.22063691 0.673591737 0.21986954  0.601774286
#>    41   73 0.845428729 0.81629794 0.950077790 0.89986858  0.849963181
#>    73   33 0.950077790 0.89986858 0.747709447 0.90900816  0.849747099
#>    33   81 0.747709447 0.90900816 0.685786240 0.90182192  0.718156312
#>    81   31 0.685786240 0.90182192 0.741666986 0.69830826  0.703750982
#>    31   41 0.741666986 0.69830826 0.845428729 0.81629794  0.750782708
#>    29   70 0.892295901 0.37562867 0.836667421 0.37329465  0.865925379
#>    40   58 1.000000000 0.38625161 0.942297742 0.67316997  1.355924671
#>    58   26 0.942297742 0.67316997 0.834439697 0.61875828  0.935243877
#>    26   84 0.834439697 0.61875828 0.802696994 0.59521455  0.886353145
#>    84   43 0.802696994 0.59521455 0.794290738 0.42310589  0.876882545
#>    43   28 0.794290738 0.42310589 0.840458680 0.39265788  0.877437401
#>    28   29 0.840458680 0.39265788 0.892295901 0.37562867  0.891152612
#>    80   39 0.240718734 0.96686189 0.283128946 0.96473448  0.261966835
#>    39   32 0.283128946 0.96473448 0.347279441 0.95794662  0.326036627
#>    32   62 0.347279441 0.95794662 0.447773761 1.00000000  0.340704732
#>    62   74 0.447773761 1.00000000 0.158197512 0.99102621  0.269953977
#>    74   54 0.158197512 0.99102621 0.006315323 0.93715980  0.061912393
#>    54   76 0.006315323 0.93715980 0.025429672 0.77741184 -0.014798987
#>    76   44 0.025429672 0.77741184 0.110211316 0.85610656  0.036573360
#>    54   83 0.006315323 0.93715980 0.022419730 0.69681308 -0.755262479
#>    99   30 0.654417496 0.40436483 0.641509881 0.44241556  0.503157668
#>    51   66 0.200660136 0.27178338 0.215557434 0.33274457  0.198633673
#>    91   88 0.041737815 0.35170314 0.038138308 0.28130555  0.108037844
#>    58   41 0.942297742 0.67316997 0.845428729 0.81629794  0.852622561
#>    31   84 0.741666986 0.69830826 0.802696994 0.59521455  0.783915210
#>    94   45 0.114516864 0.57436537 0.103152743 0.70658208  0.095337829
#>    83   46 0.022419730 0.69681308 0.010056833 0.45758950 -0.107874805
#>    86   75 0.673476387 0.39050246 0.668376406 0.50457353  0.710864694
#>    81   36 0.685786240 0.90182192 0.737461034 0.69606047  0.688311055
#>    36   84 0.737461034 0.69606047 0.802696994 0.59521455  0.767874368
#>    39   79 0.283128946 0.96473448 0.245192825 0.93081229  0.308519587
#>    79   53 0.245192825 0.93081229 0.351128261 0.83867127  0.308827843
#>    53   62 0.351128261 0.83867127 0.447773761 1.00000000  0.430188285
#>    73   37 0.950077790 0.89986858 0.971261501 0.93852315  0.852243712
#>    37   62 0.971261501 0.93852315 0.447773761 1.00000000  0.777288302
#>    62   81 0.447773761 1.00000000 0.685786240 0.90182192  0.679699186
#>    92   79 0.342701871 0.83216746 0.245192825 0.93081229  0.258102051
#>    44   38 0.110211316 0.85610656 0.128040002 0.70959266  0.125802156
#>    78   55 0.614844107 0.12126683 0.529143180 0.22063691  0.588464076
#>    60   57 0.467487930 0.09772376 0.490619193 0.00000000  0.529028343
#>    81   95 0.685786240 0.90182192 0.627437980 0.66958687  0.657764130
#>    95   84 0.627437980 0.66958687 0.802696994 0.59521455  0.701994728
#>    73   58 0.950077790 0.89986858 0.942297742 0.67316997  1.219909882
#>    44   45 0.110211316 0.85610656 0.103152743 0.70658208  0.106734880
#>    90   63 0.956687703 0.17536033 0.950458071 0.10850535  2.236688173
#>    84   75 0.802696994 0.59521455 0.668376406 0.50457353  0.761815427
#>    83   59 0.022419730 0.69681308 0.001605756 0.42266691 -0.717236596
#>    53   89 0.351128261 0.83867127 0.440003910 0.74627100  0.422661800
#>    63   61 0.950458071 0.10850535 0.910825034 0.09108609  0.959884991
#>    61   72 0.910825034 0.09108609 0.843603004 0.12378220  0.856791549
#>    72   65 0.843603004 0.12378220 0.789142238 0.10541048  0.835728575
#>    93   68 0.381749537 0.07421042 0.199724353 0.04448306  0.321779462
#>    85   91 0.000000000 0.23317149 0.041737815 0.35170314 -0.081886379
#>     1   85 0.022823304 0.22203051 0.000000000 0.23317149  0.023979680
#>     1   52 0.022823304 0.22203051 0.009474413 0.17004206 -0.001752625
#>     1   96 0.022823304 0.22203051 0.036748940 0.15072695  0.043403487
#>     1   15 0.022823304 0.22203051 0.129978413 0.21895214  0.075678350
#>     1   88 0.022823304 0.22203051 0.038138308 0.28130555  0.076951577
#>     2  100 0.812178852 0.37330070 0.809837702 0.39360708  0.824426024
#>     2   17 0.812178852 0.37330070 0.701320441 0.33106643  0.747620619
#>     2   42 0.812178852 0.37330070 0.809663713 0.33765366  0.753963825
#>     2   70 0.812178852 0.37330070 0.836667421 0.37329465  0.824418505
#>     3   75 0.446313894 0.67020786 0.668376406 0.50457353  0.516896719
#>     3   95 0.446313894 0.67020786 0.627437980 0.66958687  0.536497231
#>     3   49 0.446313894 0.67020786 0.472123903 0.72748893  0.536855260
#>     3   89 0.446313894 0.67020786 0.440003910 0.74627100  0.439120541
#>     3   64 0.446313894 0.67020786 0.403310175 0.74004239  0.427800379
#>     3    5 0.446313894 0.67020786 0.288583668 0.68598089  0.366565037
#>     3    4 0.446313894 0.67020786 0.300162730 0.59947505  0.364855566
#>     3   14 0.446313894 0.67020786 0.281220730 0.53113595  0.424894273
#>     3    9 0.446313894 0.67020786 0.314380451 0.37780051  0.457782908
#>     3   30 0.446313894 0.67020786 0.641509881 0.44241556  0.462796194
#>     4   94 0.300162730 0.59947505 0.114516864 0.57436537  0.201473369
#>     4   97 0.300162730 0.59947505 0.271888992 0.53096363  0.205796716
#>     5   92 0.288583668 0.68598089 0.342701871 0.83216746  0.326058237
#>     5   34 0.288583668 0.68598089 0.199903247 0.78310528  0.283957227
#>     5   38 0.288583668 0.68598089 0.128040002 0.70959266  0.208982648
#>     6   74 0.138776511 0.93405805 0.158197512 0.99102621  0.190464710
#>     6   22 0.138776511 0.93405805 0.034555363 0.92425262  0.081372185
#>     6   44 0.138776511 0.93405805 0.110211316 0.85610656  0.088635348
#>     6   79 0.138776511 0.93405805 0.245192825 0.93081229  0.190111975
#>     6   80 0.138776511 0.93405805 0.240718734 0.96686189  0.192293158
#>     7   71 0.335283904 0.11711821 0.255938815 0.09020146  0.274315866
#>     7   87 0.335283904 0.11711821 0.254336020 0.08170851  0.304796729
#>     7   93 0.335283904 0.11711821 0.381749537 0.07421042  0.316400710
#>     7   12 0.335283904 0.11711821 0.429521324 0.11991507  0.382315943
#>     7   69 0.335283904 0.11711821 0.360440337 0.25153041  0.380631539
#>     7   77 0.335283904 0.11711821 0.197849803 0.18381704  0.288399097
#>     8   94 0.111988229 0.45317693 0.114516864 0.57436537  0.128420008
#>     8   67 0.111988229 0.45317693 0.060594719 0.53510635  0.117444593
#>     8   46 0.111988229 0.45317693 0.010056833 0.45758950  0.062041865
#>     8   59 0.111988229 0.45317693 0.001605756 0.42266691  0.059796815
#>     8   91 0.111988229 0.45317693 0.041737815 0.35170314  0.064176708
#>     8   25 0.111988229 0.45317693 0.182150030 0.33300439  0.116308957
#>     8   98 0.111988229 0.45317693 0.150207681 0.45489951  0.134183974
#>     9   98 0.314380451 0.37780051 0.150207681 0.45489951  0.243716225
#>     9   66 0.314380451 0.37780051 0.215557434 0.33274457  0.234743833
#>     9   69 0.314380451 0.37780051 0.360440337 0.25153041  0.291172533
#>     9   24 0.314380451 0.37780051 0.448709294 0.22997522  0.430957076
#>     9   97 0.314380451 0.37780051 0.271888992 0.53096363  0.278047937
#>    10   90 0.951471460 0.22572131 0.956687703 0.17536033  0.880379755
#>    10   40 0.951471460 0.22572131 1.000000000 0.38625161  1.230005327
#>    10   48 0.951471460 0.22572131 0.982352296 0.36026156  1.182136379
#>    10   29 0.951471460 0.22572131 0.892295901 0.37562867  0.926131326
#>    10   19 0.951471460 0.22572131 0.824561414 0.32423770  0.901812693
#>    10   56 0.951471460 0.22572131 0.758204331 0.20620359  0.853383424
#>    10   72 0.951471460 0.22572131 0.843603004 0.12378220  0.854388745
#>    11   23 0.027612929 0.75447454 0.025348304 0.75345982  0.021335658
#>    11   83 0.027612929 0.75447454 0.022419730 0.69681308  0.039766802
#>    11   45 0.027612929 0.75447454 0.103152743 0.70658208  0.060272878
#>    11   76 0.027612929 0.75447454 0.025429672 0.77741184  0.091773738
#>    12   60 0.429521324 0.11991507 0.467487930 0.09772376  0.427981189
#>    12   55 0.429521324 0.11991507 0.529143180 0.22063691  0.482559817
#>    12   24 0.429521324 0.11991507 0.448709294 0.22997522  0.482207156
#>    13   50 0.660576184 0.10809088 0.886639057 0.01361826  0.693372848
#>    13   65 0.660576184 0.10809088 0.789142238 0.10541048  0.721292455
#>    13   56 0.660576184 0.10809088 0.758204331 0.20620359  0.725574156
#>    13   82 0.660576184 0.10809088 0.673591737 0.21986954  0.707220025
#>    13   78 0.660576184 0.10809088 0.614844107 0.12126683  0.652406836
#>    13   57 0.660576184 0.10809088 0.490619193 0.00000000  0.606322286
#>    15   77 0.129978413 0.21895214 0.197849803 0.18381704  0.125524670
#>    15   51 0.129978413 0.21895214 0.200660136 0.27178338  0.177941223
#>    15   25 0.129978413 0.21895214 0.182150030 0.33300439  0.135360536
#>    16   82 0.747208615 0.28061023 0.673591737 0.21986954  0.690136085
#>    16   56 0.747208615 0.28061023 0.758204331 0.20620359  0.720023121
#>    16   19 0.747208615 0.28061023 0.824561414 0.32423770  0.814057516
#>    16   42 0.747208615 0.28061023 0.809663713 0.33765366  0.788019521
#>    16   17 0.747208615 0.28061023 0.701320441 0.33106643  0.755499068
#>    17   43 0.701320441 0.33106643 0.794290738 0.42310589  0.746065997
#>    17   86 0.701320441 0.33106643 0.673476387 0.39050246  0.739691550
#>    17   99 0.701320441 0.33106643 0.654417496 0.40436483  0.611391965
#>    17   55 0.701320441 0.33106643 0.529143180 0.22063691  0.591683897
#>    18   73 0.751608109 0.89194008 0.950077790 0.89986858  0.849963181
#>    18   33 0.751608109 0.89194008 0.747709447 0.90900816  0.849747099
#>    18   81 0.751608109 0.89194008 0.685786240 0.90182192  0.718156312
#>    18   31 0.751608109 0.89194008 0.741666986 0.69830826  0.703750982
#>    18   41 0.751608109 0.89194008 0.845428729 0.81629794  0.750782708
#>    19   70 0.824561414 0.32423770 0.836667421 0.37329465  0.865925379
#>    20   40 0.972433896 0.43860521 1.000000000 0.38625161  0.956517382
#>    20   58 0.972433896 0.43860521 0.942297742 0.67316997  1.355924671
#>    20   26 0.972433896 0.43860521 0.834439697 0.61875828  0.935243877
#>    20   84 0.972433896 0.43860521 0.802696994 0.59521455  0.886353145
#>    20   43 0.972433896 0.43860521 0.794290738 0.42310589  0.876882545
#>    20   28 0.972433896 0.43860521 0.840458680 0.39265788  0.877437401
#>    20   29 0.972433896 0.43860521 0.892295901 0.37562867  0.891152612
#>    20   48 0.972433896 0.43860521 0.982352296 0.36026156  0.941931131
#>    21   80 0.251883602 0.98535965 0.240718734 0.96686189  0.205788768
#>    21   39 0.251883602 0.98535965 0.283128946 0.96473448  0.261966835
#>    21   32 0.251883602 0.98535965 0.347279441 0.95794662  0.326036627
#>    21   62 0.251883602 0.98535965 0.447773761 1.00000000  0.340704732
#>    21   74 0.251883602 0.98535965 0.158197512 0.99102621  0.269953977
#>    22   54 0.034555363 0.92425262 0.006315323 0.93715980  0.061912393
#>    22   76 0.034555363 0.92425262 0.025429672 0.77741184 -0.014798987
#>    23   54 0.025348304 0.75345982 0.006315323 0.93715980 -0.730307596
#>    24   99 0.448709294 0.22997522 0.654417496 0.40436483  0.505851763
#>    25   66 0.182150030 0.33300439 0.215557434 0.33274457  0.198633673
#>    26   41 0.834439697 0.61875828 0.845428729 0.81629794  0.852622561
#>    26   31 0.834439697 0.61875828 0.741666986 0.69830826  0.838698684
#>    27   45 0.064660774 0.57893955 0.103152743 0.70658208  0.095337829
#>    27   83 0.064660774 0.57893955 0.022419730 0.69681308  0.069388025
#>    27   46 0.064660774 0.57893955 0.010056833 0.45758950 -0.107874805
#>    27   67 0.064660774 0.57893955 0.060594719 0.53510635 -0.077706843
#>    27   94 0.064660774 0.57893955 0.114516864 0.57436537  0.087575537
#>    28  100 0.840458680 0.39265788 0.809837702 0.39360708  0.826011345
#>    28   70 0.840458680 0.39265788 0.836667421 0.37329465  0.824916192
#>    30   86 0.641509881 0.44241556 0.673476387 0.39050246  0.694245389
#>    31   36 0.741666986 0.69830826 0.737461034 0.69606047  0.688311055
#>    32   79 0.347279441 0.95794662 0.245192825 0.93081229  0.308519587
#>    32   53 0.347279441 0.95794662 0.351128261 0.83867127  0.308827843
#>    33   37 0.747709447 0.90900816 0.971261501 0.93852315  0.852243712
#>    33   62 0.747709447 0.90900816 0.447773761 1.00000000  0.777288302
#>    34   79 0.199903247 0.78310528 0.245192825 0.93081229  0.258102051
#>    34   44 0.199903247 0.78310528 0.110211316 0.85610656  0.192866511
#>    35   55 0.546256489 0.13041786 0.529143180 0.22063691  0.588464076
#>    35   60 0.546256489 0.13041786 0.467487930 0.09772376  0.485475457
#>    35   57 0.546256489 0.13041786 0.490619193 0.00000000  0.529028343
#>    35   78 0.546256489 0.13041786 0.614844107 0.12126683  0.569551228
#>    36   95 0.737461034 0.69606047 0.627437980 0.66958687  0.657764130
#>    37   58 0.971261501 0.93852315 0.942297742 0.67316997  1.219909882
#>    37   40 0.971261501 0.93852315 1.000000000 0.38625161  1.856101726
#>    38   45 0.128040002 0.70959266 0.103152743 0.70658208  0.106734880
#>    40   63 1.000000000 0.38625161 0.950458071 0.10850535  2.236688173
#>    43   75 0.794290738 0.42310589 0.668376406 0.50457353  0.761815427
#>    47   49 0.496836119 0.82308920 0.472123903 0.72748893  0.479305645
#>    47   95 0.496836119 0.82308920 0.627437980 0.66958687  0.570139379
#>    47   81 0.496836119 0.82308920 0.685786240 0.90182192  0.619396080
#>    47   62 0.496836119 0.82308920 0.447773761 1.00000000  0.560647737
#>    47   53 0.496836119 0.82308920 0.351128261 0.83867127  0.431395189
#>    47   89 0.496836119 0.82308920 0.440003910 0.74627100  0.422661800
#>    50   61 0.886639057 0.01361826 0.910825034 0.09108609  0.959884991
#>    50   72 0.886639057 0.01361826 0.843603004 0.12378220  0.856791549
#>    51   69 0.200660136 0.27178338 0.360440337 0.25153041  0.275948153
#>    53   92 0.351128261 0.83867127 0.342701871 0.83216746  0.303822126
#>    53   64 0.351128261 0.83867127 0.403310175 0.74004239  0.380946012
#>    57   93 0.490619193 0.00000000 0.381749537 0.07421042  0.437492549
#>    57   68 0.490619193 0.00000000 0.199724353 0.04448306  0.321779462
#>    61   90 0.910825034 0.09108609 0.956687703 0.17536033  0.910351611
#>    68   87 0.199724353 0.04448306 0.254336020 0.08170851  0.310004587
#>    68   71 0.199724353 0.04448306 0.255938815 0.09020146  0.204998642
#>    68   77 0.199724353 0.04448306 0.197849803 0.18381704  0.189861161
#>    68   96 0.199724353 0.04448306 0.036748940 0.15072695  0.128405103
#>    50   63 0.886639057 0.01361826 0.950458071 0.10850535  0.959884991
#>    54   59 0.006315323 0.93715980 0.001605756 0.42266691 -1.786718262
#>    59   85 0.001605756 0.42266691 0.000000000 0.23317149 -0.081886379
#>           my1           mx2           my2 bp1 bp2
#>   0.253347740  -0.062329328   0.321733394   0   0
#>   0.200632788 -15.493298485  -2.124325801   0   1
#>   0.189038211 -17.248962302 -24.229202605   0   1
#>   0.195341515   0.125524670   0.127226699   0   0
#>   0.239661292   0.114512961   0.294985310   0   0
#>   0.385000839   0.824916192   0.385648279   0   0
#>   0.376145844   0.746065997   0.378843342   0   0
#>   0.359495912   0.755499068   0.334245069   0   0
#>   0.354524860   0.832697655   0.348252089   0   0
#>   0.533162323   0.710864694   0.449323589   0   0
#>   0.559440323   0.701475014   0.600369929   0   0
#>   0.663866586   0.570139379   0.753146527   0   0
#>   0.707904418   0.479305645   0.776626614   0   0
#>   0.706965329   0.410799022   0.807123008   0   0
#>   0.669256942   0.326058237   0.755218376   0   0
#>   0.652162243   0.201473369   0.630293055   0   0
#>   0.528107694   0.275846292   0.569420310   0   0
#>   0.489065382   0.278047937   0.450196603   0   0
#>   0.486803401   0.481063545   0.394320486   0   0
#>   0.630293055   0.198961427   0.634210483   0   0
#>   0.598328824   0.187235211   0.531025914   0   0
#>   0.569420310   0.278047937   0.450196603   0   0
#>   0.755218376   0.380946012   0.791328576   0   0
#>   0.770804151   0.258102051   0.846057293   0   0
#>   0.702347839   0.125802156   0.783662047   0   0
#>   0.634210483   0.123621759   0.641744677   0   0
#>   0.948231526   0.205788768   1.000563219   0   0
#>   0.985422230   0.061912393   1.021455196   0   0
#>   0.908222618   0.036573360   0.850423251   0   0
#>   0.871036676   0.192866511   0.866059662   0   0
#>   0.942549383   0.261186524   0.951099692   0   0
#>   0.166434601   0.189861161   0.114029968   0   0
#>   0.076583240   0.204998642   0.095417211   0   0
#>   0.050056111   0.310004587  -0.058631982   0   0
#>   0.121436935   0.427981189   0.073706331   0   0
#>   0.178191224   0.390571352   0.183408337   0   0
#>   0.195453341   0.275948153   0.225350033   0   0
#>   0.513454678   0.187235211   0.531025914   0   0
#>   0.513683684   0.087575537   0.554708741   0   0
#>   0.478930102  -0.077706843   0.570040665   0   0
#>   0.427068839  -0.717236596   0.615106723   0   0
#>   0.411222785  -0.081886379   0.328619897   0   0
#>   0.375131574   0.108037844   0.313022329   0   0
#>   0.385567763   0.199396449   0.402656548   0   0
#>   0.440672087   0.187235211   0.531025914   0   0
#>   0.421566489   0.199396449   0.402656548   0   0
#>   0.297799154   0.283366248   0.283873045   0   0
#>   0.348788738   0.390571352   0.183408337   0   0
#>   0.394320486   0.503157668   0.374268928   0   0
#>   0.192907208   0.898966365   0.152156189   0   0
#>   0.229120400   2.236688173   0.022370526   0   0
#>   0.243591251   0.956517382   0.396790493   0   0
#>   0.302351738   0.941931131   0.394943857   0   0
#>   0.292752018   0.865925379   0.340052332   0   0
#>   0.230364829   0.814057516   0.252473289   0   0
#>   0.220410000   0.799768636   0.163816866   0   0
#>   0.765449599  -0.730307596   0.768003016   0   0
#>   0.724315326  -0.755262479   0.765417464   0   0
#>   0.722468476   0.069388025   0.647139093   0   0
#>   0.772154155   0.100418288   0.781640013   0   0
#>   0.073706331   0.437492549   0.039024370   0   0
#>   0.167083672   0.485475457   0.165621139   0   0
#>   0.167432482   0.505851763   0.371091000   0   0
#>  -0.131138665   0.780103858  -2.653282037   0   1
#>  -0.064330032   0.835728575   0.057217932   0   0
#>   0.141043266   0.799768636   0.163816866   0   0
#>   0.159306748   0.720023121   0.238577009   0   0
#>   0.165689218   0.601239544   0.196174760   0   0
#>   0.005735531   0.569551228   0.043403552   0   0
#>   0.127226699   0.128405103   0.113203160   0   0
#>   0.228481141   0.275948153   0.225350033   0   0
#>   0.285448876   0.198633673   0.304579441   0   0
#>   0.294985310   0.108037844   0.313022329   0   0
#>   0.274799684   0.601774286   0.296834098   0   0
#>   0.298639424   0.832697655   0.348252089   0   0
#>   0.378843342   0.826011345   0.420977507   0   0
#>   0.385282257   0.722271483   0.449833573   0   0
#>   0.325177630   0.694245389   0.439089928   0   0
#>   0.312566625   0.505851763   0.371091000   0   0
#>   0.296834098   0.601239544   0.196174760   0   0
#>   0.917927084   0.955148404   0.786211758   0   0
#>   0.923336137   0.852243712   0.978616113   0   0
#>   0.893278376   0.679699186   1.224660201   0   0
#>   0.797325980   0.688311055   0.793086488   0   0
#>   0.794911356   0.838698684   0.717596843   0   0
#>   0.340052332   0.864336140   0.377929975   0   0
#>   0.607093197   1.856101726   0.707683993   0   0
#>   0.553045431   0.852622561   0.716822265   0   0
#>   0.515595960   0.783915210   0.653707301   0   0
#>   0.505331500   0.761815427   0.510951691   0   0
#>   0.498954208   0.826011345   0.420977507   0   0
#>   0.459559799   0.864336140   0.377929975   0   0
#>   0.966655293   0.261186524   0.951099692   0   0
#>   1.063715462   0.308519587   0.898165822   0   0
#>   1.114759669   0.430188285   0.900922181   0   0
#>   2.061415677   0.163040348   5.511424380   0   1
#>   1.021455196  -2.012282626   6.869873470   0   1
#>   0.853615880  -0.730307596   0.768003016   0   0
#>   0.850423251   0.100418288   0.781640013   0   0
#>   0.765417464  -1.786718262   0.696304881   0   0
#>   0.374268928   0.694245389   0.439089928   0   0
#>   0.304579441   0.283366248   0.283873045   0   0
#>   0.313022329  -0.062329328   0.321733394   0   0
#>   0.716822265   0.955148404   0.786211758   0   0
#>   0.653707301   0.767874368   0.644211350   0   0
#>   0.639313653   0.123621759   0.641744677   0   0
#>   0.583615358  -0.717236596   0.615106723   0   0
#>   0.449323589   0.722271483   0.449833573   0   0
#>   0.793086488   0.657764130   0.785414953   0   0
#>   0.644211350   0.701994728   0.601594639   0   0
#>   0.898165822   0.261186524   0.951099692   0   0
#>   0.897006079   0.303822126   0.891250956   0   0
#>   0.900922181   0.431395189   0.900199173   0   0
#>   0.978616113   1.219909882   0.777125434   0   0
#>   1.546342425   0.998572284   3.430619900   0   1
#>   1.224660201   0.560647737   0.936044555   0   0
#>   0.846057293   0.303822126   0.891250956   0   0
#>   0.783662047   0.106734880   0.781341827   0   0
#>   0.185156660   0.601239544   0.196174760   0   0
#>   0.060690937   0.437492549   0.039024370   0   0
#>   0.785414953   0.619396080   0.795054794   0   0
#>   0.601594639   0.701475014   0.600369929   0   0
#>   0.777125434   0.955148404   0.786211758   0   0
#>   0.781341827   0.100418288   0.781640013   0   0
#>   0.022370526   0.910351611   0.145960253   0   0
#>   0.510951691   0.701475014   0.600369929   0   0
#>   0.615106723  -1.786718262   0.696304881   0   0
#>   0.818533279   0.410799022   0.807123008   0   0
#>   0.033259814   0.910351611   0.145960253   0   0
#>   0.065446280   0.898966365   0.152156189   0   0
#>   0.057217932   0.799768636   0.163816866   0   0
#>  -0.130731348   0.310004587  -0.058631982   0   0
#>   0.328619897  -0.062329328   0.321733394   0   0
#>   0.253347740  -0.001752625   0.200632788   0   0
#>   0.200632788   0.043403487   0.189038211   0   0
#>   0.189038211   0.075678350   0.195341515   0   0
#>   0.195341515   0.076951577   0.239661292   0   0
#>   0.239661292   0.023979680   0.253347740   0   0
#>   0.385000839   0.747620619   0.376145844   0   0
#>   0.376145844   0.753963825   0.359495912   0   0
#>   0.359495912   0.824418505   0.354524860   0   0
#>   0.354524860   0.824426024   0.385000839   0   0
#>   0.533162323   0.536497231   0.559440323   0   0
#>   0.559440323   0.536855260   0.663866586   0   0
#>   0.663866586   0.439120541   0.707904418   0   0
#>   0.707904418   0.427800379   0.706965329   0   0
#>   0.706965329   0.366565037   0.669256942   0   0
#>   0.669256942   0.364855566   0.652162243   0   0
#>   0.652162243   0.424894273   0.528107694   0   0
#>   0.528107694   0.457782908   0.489065382   0   0
#>   0.489065382   0.462796194   0.486803401   0   0
#>   0.486803401   0.516896719   0.533162323   0   0
#>   0.630293055   0.205796716   0.598328824   0   0
#>   0.598328824   0.275846292   0.569420310   0   0
#>   0.755218376   0.283957227   0.770804151   0   0
#>   0.770804151   0.208982648   0.702347839   0   0
#>   0.702347839   0.198961427   0.634210483   0   0
#>   0.948231526   0.081372185   0.985422230   0   0
#>   0.985422230   0.088635348   0.908222618   0   0
#>   0.908222618   0.190111975   0.871036676   0   0
#>   0.871036676   0.192293158   0.942549383   0   0
#>   0.942549383   0.190464710   0.948231526   0   0
#>   0.166434601   0.304796729   0.076583240   0   0
#>   0.076583240   0.316400710   0.050056111   0   0
#>   0.050056111   0.382315943   0.121436935   0   0
#>   0.121436935   0.380631539   0.178191224   0   0
#>   0.178191224   0.288399097   0.195453341   0   0
#>   0.195453341   0.274315866   0.166434601   0   0
#>   0.513454678   0.117444593   0.513683684   0   0
#>   0.513683684   0.062041865   0.478930102   0   0
#>   0.478930102   0.059796815   0.427068839   0   0
#>   0.427068839   0.064176708   0.411222785   0   0
#>   0.411222785   0.116308957   0.375131574   0   0
#>   0.375131574   0.134183974   0.385567763   0   0
#>   0.385567763   0.128420008   0.513454678   0   0
#>   0.440672087   0.234743833   0.421566489   0   0
#>   0.421566489   0.291172533   0.297799154   0   0
#>   0.297799154   0.430957076   0.348788738   0   0
#>   0.348788738   0.481063545   0.394320486   0   0
#>   0.450196603   0.243716225   0.440672087   0   0
#>   0.192907208   1.230005327   0.229120400   0   0
#>   0.229120400   1.182136379   0.243591251   0   0
#>   0.243591251   0.926131326   0.302351738   0   0
#>   0.302351738   0.901812693   0.292752018   0   0
#>   0.292752018   0.853383424   0.230364829   0   0
#>   0.230364829   0.854388745   0.220410000   0   0
#>   0.220410000   0.880379755   0.192907208   0   0
#>   0.765449599   0.039766802   0.724315326   0   0
#>   0.724315326   0.060272878   0.722468476   0   0
#>   0.722468476   0.091773738   0.772154155   0   0
#>   0.772154155   0.021335658   0.765449599   0   0
#>   0.073706331   0.482559817   0.167083672   0   0
#>   0.167083672   0.482207156   0.167432482   0   0
#>   0.167432482   0.390571352   0.183408337   0   0
#>  -0.131138665   0.721292455  -0.064330032   0   0
#>  -0.064330032   0.725574156   0.141043266   0   0
#>   0.141043266   0.707220025   0.159306748   0   0
#>   0.159306748   0.652406836   0.165689218   0   0
#>   0.165689218   0.606322286   0.005735531   0   0
#>   0.005735531   0.693372848  -0.131138665   0   0
#>   0.127226699   0.177941223   0.228481141   0   0
#>   0.228481141   0.135360536   0.285448876   0   0
#>   0.285448876   0.114512961   0.294985310   0   0
#>   0.274799684   0.720023121   0.238577009   0   0
#>   0.238577009   0.814057516   0.252473289   0   0
#>   0.252473289   0.788019521   0.298639424   0   0
#>   0.298639424   0.755499068   0.334245069   0   0
#>   0.334245069   0.690136085   0.274799684   0   0
#>   0.378843342   0.739691550   0.385282257   0   0
#>   0.385282257   0.611391965   0.325177630   0   0
#>   0.325177630   0.591683897   0.312566625   0   0
#>   0.312566625   0.601774286   0.296834098   0   0
#>   0.917927084   0.849747099   0.923336137   0   0
#>   0.923336137   0.718156312   0.893278376   0   0
#>   0.893278376   0.703750982   0.797325980   0   0
#>   0.797325980   0.750782708   0.794911356   0   0
#>   0.794911356   0.849963181   0.917927084   0   0
#>   0.340052332   0.832697655   0.348252089   0   0
#>   0.396790493   1.355924671   0.607093197   0   0
#>   0.607093197   0.935243877   0.553045431   0   0
#>   0.553045431   0.886353145   0.515595960   0   0
#>   0.515595960   0.876882545   0.505331500   0   0
#>   0.505331500   0.877437401   0.498954208   0   0
#>   0.498954208   0.891152612   0.459559799   0   0
#>   0.459559799   0.941931131   0.394943857   0   0
#>   0.394943857   0.956517382   0.396790493   0   0
#>   1.000563219   0.261966835   0.966655293   0   0
#>   0.966655293   0.326036627   1.063715462   0   0
#>   1.063715462   0.340704732   1.114759669   0   0
#>   1.114759669   0.269953977   2.061415677   0   0
#>   2.061415677   0.205788768   1.000563219   0   0
#>   1.021455196  -0.014798987   0.853615880   0   0
#>   0.853615880   0.036573360   0.850423251   0   0
#>   0.768003016  -0.755262479   0.765417464   0   0
#>   0.371091000   0.503157668   0.374268928   0   0
#>   0.304579441   0.199396449   0.402656548   0   0
#>   0.716822265   0.838698684   0.717596843   0   0
#>   0.717596843   0.783915210   0.653707301   0   0
#>   0.639313653   0.069388025   0.647139093   0   0
#>   0.647139093  -0.107874805   0.583615358   0   0
#>   0.583615358  -0.077706843   0.570040665   0   0
#>   0.570040665   0.087575537   0.554708741   0   0
#>   0.554708741   0.095337829   0.639313653   0   0
#>   0.420977507   0.824916192   0.385648279   0   0
#>   0.385648279   0.864336140   0.377929975   0   0
#>   0.439089928   0.710864694   0.449323589   0   0
#>   0.793086488   0.767874368   0.644211350   0   0
#>   0.898165822   0.308827843   0.897006079   0   0
#>   0.897006079   0.430188285   0.900922181   0   0
#>   0.978616113   0.777288302   1.546342425   0   0
#>   1.546342425   0.679699186   1.224660201   0   0
#>   0.846057293   0.192866511   0.866059662   0   0
#>   0.866059662   0.125802156   0.783662047   0   0
#>   0.185156660   0.485475457   0.165621139   0   0
#>   0.165621139   0.529028343   0.060690937   0   0
#>   0.060690937   0.569551228   0.043403552   0   0
#>   0.043403552   0.588464076   0.185156660   0   0
#>   0.785414953   0.701994728   0.601594639   0   0
#>   0.777125434   1.856101726   0.707683993   0   0
#>   0.707683993   3.661915383   0.801652940   0   1
#>   0.781341827   0.123621759   0.641744677   0   0
#>   0.022370526   5.726077266  -0.600035928   0   1
#>   0.510951691   0.722271483   0.449833573   0   0
#>   0.776626614   0.570139379   0.753146527   0   0
#>   0.753146527   0.619396080   0.795054794   0   0
#>   0.795054794   0.560647737   0.936044555   0   0
#>   0.936044555   0.431395189   0.900199173   0   0
#>   0.900199173   0.422661800   0.818533279   0   0
#>   0.818533279   0.479305645   0.776626614   0   0
#>   0.033259814   0.856791549   0.065446280   0   0
#>   0.065446280   0.835728575   0.057217932   0   0
#>   0.225350033   0.283366248   0.283873045   0   0
#>   0.891250956   0.380946012   0.791328576   0   0
#>   0.791328576   0.410799022   0.807123008   0   0
#>   0.039024370   0.321779462  -0.130731348   0   0
#>  -0.130731348  -0.191890387  -3.489850381   0   1
#>   0.145960253   0.898966365   0.152156189   0   0
#>  -0.058631982   0.204998642   0.095417211   0   0
#>   0.095417211   0.189861161   0.114029968   0   0
#>   0.114029968   0.128405103   0.113203160   0   0
#>   0.113203160  -2.678659590  -4.192762541   0   1
#>   0.033259814   8.216232129  -4.847203017   0   1
#>   0.696304881  -3.730216856   0.714095286   0   1
#>   0.328619897  -5.358679816   0.373334660   0   1
#> 
#> $x
#>               [,1]       [,2]
#>   [1,] 0.022823304 0.22203051
#>   [2,] 0.812178852 0.37330070
#>   [3,] 0.446313894 0.67020786
#>   [4,] 0.300162730 0.59947505
#>   [5,] 0.288583668 0.68598089
#>   [6,] 0.138776511 0.93405805
#>   [7,] 0.335283904 0.11711821
#>   [8,] 0.111988229 0.45317693
#>   [9,] 0.314380451 0.37780051
#>  [10,] 0.951471460 0.22572131
#>  [11,] 0.027612929 0.75447454
#>  [12,] 0.429521324 0.11991507
#>  [13,] 0.660576184 0.10809088
#>  [14,] 0.281220730 0.53113595
#>  [15,] 0.129978413 0.21895214
#>  [16,] 0.747208615 0.28061023
#>  [17,] 0.701320441 0.33106643
#>  [18,] 0.751608109 0.89194008
#>  [19,] 0.824561414 0.32423770
#>  [20,] 0.972433896 0.43860521
#>  [21,] 0.251883602 0.98535965
#>  [22,] 0.034555363 0.92425262
#>  [23,] 0.025348304 0.75345982
#>  [24,] 0.448709294 0.22997522
#>  [25,] 0.182150030 0.33300439
#>  [26,] 0.834439697 0.61875828
#>  [27,] 0.064660774 0.57893955
#>  [28,] 0.840458680 0.39265788
#>  [29,] 0.892295901 0.37562867
#>  [30,] 0.641509881 0.44241556
#>  [31,] 0.741666986 0.69830826
#>  [32,] 0.347279441 0.95794662
#>  [33,] 0.747709447 0.90900816
#>  [34,] 0.199903247 0.78310528
#>  [35,] 0.546256489 0.13041786
#>  [36,] 0.737461034 0.69606047
#>  [37,] 0.971261501 0.93852315
#>  [38,] 0.128040002 0.70959266
#>  [39,] 0.283128946 0.96473448
#>  [40,] 1.000000000 0.38625161
#>  [41,] 0.845428729 0.81629794
#>  [42,] 0.809663713 0.33765366
#>  [43,] 0.794290738 0.42310589
#>  [44,] 0.110211316 0.85610656
#>  [45,] 0.103152743 0.70658208
#>  [46,] 0.010056833 0.45758950
#>  [47,] 0.496836119 0.82308920
#>  [48,] 0.982352296 0.36026156
#>  [49,] 0.472123903 0.72748893
#>  [50,] 0.886639057 0.01361826
#>  [51,] 0.200660136 0.27178338
#>  [52,] 0.009474413 0.17004206
#>  [53,] 0.351128261 0.83867127
#>  [54,] 0.006315323 0.93715980
#>  [55,] 0.529143180 0.22063691
#>  [56,] 0.758204331 0.20620359
#>  [57,] 0.490619193 0.00000000
#>  [58,] 0.942297742 0.67316997
#>  [59,] 0.001605756 0.42266691
#>  [60,] 0.467487930 0.09772376
#>  [61,] 0.910825034 0.09108609
#>  [62,] 0.447773761 1.00000000
#>  [63,] 0.950458071 0.10850535
#>  [64,] 0.403310175 0.74004239
#>  [65,] 0.789142238 0.10541048
#>  [66,] 0.215557434 0.33274457
#>  [67,] 0.060594719 0.53510635
#>  [68,] 0.199724353 0.04448306
#>  [69,] 0.360440337 0.25153041
#>  [70,] 0.836667421 0.37329465
#>  [71,] 0.255938815 0.09020146
#>  [72,] 0.843603004 0.12378220
#>  [73,] 0.950077790 0.89986858
#>  [74,] 0.158197512 0.99102621
#>  [75,] 0.668376406 0.50457353
#>  [76,] 0.025429672 0.77741184
#>  [77,] 0.197849803 0.18381704
#>  [78,] 0.614844107 0.12126683
#>  [79,] 0.245192825 0.93081229
#>  [80,] 0.240718734 0.96686189
#>  [81,] 0.685786240 0.90182192
#>  [82,] 0.673591737 0.21986954
#>  [83,] 0.022419730 0.69681308
#>  [84,] 0.802696994 0.59521455
#>  [85,] 0.000000000 0.23317149
#>  [86,] 0.673476387 0.39050246
#>  [87,] 0.254336020 0.08170851
#>  [88,] 0.038138308 0.28130555
#>  [89,] 0.440003910 0.74627100
#>  [90,] 0.956687703 0.17536033
#>  [91,] 0.041737815 0.35170314
#>  [92,] 0.342701871 0.83216746
#>  [93,] 0.381749537 0.07421042
#>  [94,] 0.114516864 0.57436537
#>  [95,] 0.627437980 0.66958687
#>  [96,] 0.036748940 0.15072695
#>  [97,] 0.271888992 0.53096363
#>  [98,] 0.150207681 0.45489951
#>  [99,] 0.654417496 0.40436483
#> [100,] 0.809837702 0.39360708
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.0228233,0.2220305) [5]: 15 52 85 88 96 
#> 2: (0.8121789,0.3733007) [4]: 17 42 70 100 
#> 3: (0.4463139,0.6702079) [10]: 4 5 9 14 30 49 64 75 89 95 
#> 4: (0.3001627,0.599475) [5]: 3 5 14 94 97 
#> 5: (0.2885837,0.6859809) [7]: 3 4 34 38 64 92 94 
#> 6: (0.1387765,0.9340581) [5]: 22 44 74 79 80 
#> 7: (0.3352839,0.1171182) [6]: 12 69 71 77 87 93 
#> 8: (0.1119882,0.4531769) [7]: 25 46 59 67 91 94 98 
#> 9: (0.3143805,0.3778005) [8]: 3 14 24 30 66 69 97 98 
#> 10: (0.9514715,0.2257213) [7]: 19 29 40 48 56 72 90 
#> 11: (0.02761293,0.7544745) [4]: 23 45 76 83 
#> 12: (0.4295213,0.1199151) [6]: 7 24 55 60 69 93 
#> 13: (0.6605762,0.1080909) [6]: 50 56 57 65 78 82 
#> 14: (0.2812207,0.531136) [4]: 3 4 9 97 
#> 15: (0.1299784,0.2189521) [6]: 1 25 51 77 88 96 
#> 16: (0.7472086,0.2806102) [5]: 17 19 42 56 82 
#> 17: (0.7013204,0.3310664) [9]: 2 16 42 43 55 82 86 99 100 
#> 18: (0.7516081,0.8919401) [5]: 31 33 41 73 81 
#> 19: (0.8245614,0.3242377) [6]: 10 16 29 42 56 70 
#> 20: (0.9724339,0.4386052) [8]: 26 28 29 40 43 48 58 84 
#> 21: (0.2518836,0.9853596) [5]: 32 39 62 74 80 
#> 22: (0.03455536,0.9242526) [5]: 6 44 54 74 76 
#> 23: (0.0253483,0.7534598) [4]: 11 54 76 83 
#> 24: (0.4487093,0.2299752) [6]: 9 12 30 55 69 99 
#> 25: (0.18215,0.3330044) [7]: 8 15 51 66 88 91 98 
#> 26: (0.8344397,0.6187583) [5]: 20 31 41 58 84 
#> 27: (0.06466077,0.5789395) [5]: 45 46 67 83 94 
#> 28: (0.8404587,0.3926579) [5]: 20 29 43 70 100 
#> 29: (0.8922959,0.3756287) [6]: 10 19 20 28 48 70 
#> 30: (0.6415099,0.4424156) [6]: 3 9 24 75 86 99 
#> 31: (0.741667,0.6983083) [6]: 18 26 36 41 81 84 
#> 32: (0.3472794,0.9579466) [5]: 21 39 53 62 79 
#> 33: (0.7477094,0.9090082) [5]: 18 37 62 73 81 
#> 34: (0.1999032,0.7831053) [5]: 5 38 44 79 92 
#> 35: (0.5462565,0.1304179) [4]: 55 57 60 78 
#> 36: (0.737461,0.6960605) [4]: 31 81 84 95 
#> 37: (0.9712615,0.9385231) [5]: 33 40 58 62 73 
#> 38: (0.12804,0.7095927) [5]: 5 34 44 45 94 
#> 39: (0.2831289,0.9647345) [4]: 21 32 79 80 
#> 40: (1,0.3862516) [7]: 10 20 37 48 58 63 90 
#> 41: (0.8454287,0.8162979) [5]: 18 26 31 58 73 
#> 42: (0.8096637,0.3376537) [5]: 2 16 17 19 70 
#> 43: (0.7942907,0.4231059) [7]: 17 20 28 75 84 86 100 
#> 44: (0.1102113,0.8561066) [7]: 6 22 34 38 45 76 79 
#> 45: (0.1031527,0.7065821) [7]: 11 27 38 44 76 83 94 
#> 46: (0.01005683,0.4575895) [5]: 8 27 59 67 83 
#> 47: (0.4968361,0.8230892) [6]: 49 53 62 81 89 95 
#> 48: (0.9823523,0.3602616) [4]: 10 20 29 40 
#> 49: (0.4721239,0.7274889) [4]: 3 47 89 95 
#> 50: (0.8866391,0.01361826) [6]: 13 57 61 63 65 72 
#> 51: (0.2006601,0.2717834) [5]: 15 25 66 69 77 
#> 52: (0.009474413,0.1700421) [3]: 1 85 96 
#> 53: (0.3511283,0.8386713) [7]: 32 47 62 64 79 89 92 
#> 54: (0.006315323,0.9371598) [6]: 22 23 59 74 76 83 
#> 55: (0.5291432,0.2206369) [8]: 12 17 24 35 60 78 82 99 
#> 56: (0.7582043,0.2062036) [7]: 10 13 16 19 65 72 82 
#> 57: (0.4906192,0) [7]: 13 35 50 60 68 78 93 
#> 58: (0.9422977,0.67317) [6]: 20 26 37 40 41 73 
#> 59: (0.001605756,0.4226669) [6]: 8 46 54 83 85 91 
#> 60: (0.4674879,0.09772376) [5]: 12 35 55 57 93 
#> 61: (0.910825,0.09108609) [4]: 50 63 72 90 
#> 62: (0.4477738,1) [8]: 21 32 33 37 47 53 74 81 
#> 63: (0.9504581,0.1085054) [4]: 40 50 61 90 
#> 64: (0.4033102,0.7400424) [5]: 3 5 53 89 92 
#> 65: (0.7891422,0.1054105) [4]: 13 50 56 72 
#> 66: (0.2155574,0.3327446) [5]: 9 25 51 69 98 
#> 67: (0.06059472,0.5351064) [4]: 8 27 46 94 
#> 68: (0.1997244,0.04448306) [6]: 57 71 77 87 93 96 
#> 69: (0.3604403,0.2515304) [7]: 7 9 12 24 51 66 77 
#> 70: (0.8366674,0.3732947) [6]: 2 19 28 29 42 100 
#> 71: (0.2559388,0.09020146) [4]: 7 68 77 87 
#> 72: (0.843603,0.1237822) [6]: 10 50 56 61 65 90 
#> 73: (0.9500778,0.8998686) [5]: 18 33 37 41 58 
#> 74: (0.1581975,0.9910262) [6]: 6 21 22 54 62 80 
#> 75: (0.6683764,0.5045735) [6]: 3 30 43 84 86 95 
#> 76: (0.02542967,0.7774118) [6]: 11 22 23 44 45 54 
#> 77: (0.1978498,0.183817) [7]: 7 15 51 68 69 71 96 
#> 78: (0.6148441,0.1212668) [5]: 13 35 55 57 82 
#> 79: (0.2451928,0.9308123) [8]: 6 32 34 39 44 53 80 92 
#> 80: (0.2407187,0.9668619) [5]: 6 21 39 74 79 
#> 81: (0.6857862,0.9018219) [7]: 18 31 33 36 47 62 95 
#> 82: (0.6735917,0.2198695) [6]: 13 16 17 55 56 78 
#> 83: (0.02241973,0.6968131) [7]: 11 23 27 45 46 54 59 
#> 84: (0.802697,0.5952145) [7]: 20 26 31 36 43 75 95 
#> 85: (0,0.2331715) [5]: 1 52 59 88 91 
#> 86: (0.6734764,0.3905025) [5]: 17 30 43 75 99 
#> 87: (0.254336,0.08170851) [4]: 7 68 71 93 
#> 88: (0.03813831,0.2813056) [5]: 1 15 25 85 91 
#> 89: (0.4400039,0.746271) [5]: 3 47 49 53 64 
#> 90: (0.9566877,0.1753603) [5]: 10 40 61 63 72 
#> 91: (0.04173782,0.3517031) [5]: 8 25 59 85 88 
#> 92: (0.3427019,0.8321675) [5]: 5 34 53 64 79 
#> 93: (0.3817495,0.07421042) [6]: 7 12 57 60 68 87 
#> 94: (0.1145169,0.5743654) [9]: 4 5 8 27 38 45 67 97 98 
#> 95: (0.627438,0.6695869) [7]: 3 36 47 49 75 81 84 
#> 96: (0.03674894,0.150727) [5]: 1 15 52 68 77 
#> 97: (0.271889,0.5309636) [5]: 4 9 14 94 98 
#> 98: (0.1502077,0.4548995) [6]: 8 9 25 66 94 97 
#> 99: (0.6544175,0.4043648) [5]: 17 24 30 55 86 
#> 100: (0.8098377,0.3936071) [5]: 2 17 28 43 70 
#> number of nodes: 100 
#> number of arcs: 284 
#> number of boundary nodes: 13 
#> boundary nodes:  37 40 50 52 54 57 59 62 63 68 74 85 96 
#> number of triangles: 185 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.061411878 0.063836422 0.033421153 0.115526667 0.111006997 0.033651575 
#>                                                                         
#> 0.125249070 0.108543338 0.044715551 0.067715754 0.170015752 0.165756179 
#>                                                                         
#> 0.037208335 0.037218621 0.126825931 0.087277348 0.070915663 0.156879968 
#>                                                                         
#> 0.333449799 0.206778327 0.163247352 0.009333329 0.110274187 0.150991870 
#>                                                                         
#> 0.102802873 0.135901785 0.085986433 0.140520792 0.101821948 0.154275577 
#>                                                                         
#> 0.036326179 0.110173551 0.008642866 0.127633952 0.066113981 0.148643132 
#>                                                                         
#> 0.176127175 0.124683300 0.066699839 0.092536159 0.035930594 0.081525696 
#>                                                                         
#> 0.141651804 0.126010856 0.143499446 0.138536715 0.166092729 0.090862726 
#>                                                                         
#> 0.286884932 0.124291803 0.215293032 0.031415355 0.091358100 0.085023479 
#>                                                                         
#> 0.135407954 0.118685378 0.023952152 0.056722398 0.081321910 0.105155740 
#>                                                                         
#> 0.088904157 0.137510038 0.080974157 0.396253946 0.133908333 0.105434365 
#>                                                                         
#> 0.085709096 0.114777083 0.173601477 0.164464107 0.088011219 0.063958076 
#>                                                                         
#> 0.153010279 0.114602045 0.020048177 0.033344977 0.125136289 0.023567067 
#>                                                                         
#> 0.222372668 0.144450595 0.133923407 0.202574624 0.062338797 0.211046130 
#>                                                                         
#> 0.157124360 0.055677423 0.292663105 0.120805585 0.039520966 0.172313831 
#>                                                                         
#> 0.055304249 0.054562730 0.042463537 0.064508612 0.108938491 0.289715262 
#>                                                                         
#> 0.161151447 0.160887451 0.115675350 0.240885662 0.040180405 0.062755045 
#>                                                                         
#> 0.070489548 0.172827144 0.119803900 0.132704191 0.239542820 0.114185025 
#>                                                                         
#> 0.212151027 0.120106746 0.050890709 0.140400438 0.188061989 0.044078619 
#>                                                                         
#> 0.527085209 0.257466261 0.138704055 0.147594662 0.131221421 0.100424048 
#>                                                                         
#> 0.239452788 0.190386355 0.226832078 0.149690989 0.067144593 0.162042629 
#>                                                                         
#> 0.274935158 0.128205661 0.043292126 0.074751837 0.057476040 0.184436665 
#>                                                                         
#> 0.125665419 0.025397335 0.053674867 0.072650677 0.107199318 0.061221568 
#>                                                                         
#> 0.020440898 0.118631028 0.035735653 0.024488569 0.277031568 0.181125151 
#>                                                                         
#> 0.062827359 0.076324418 0.082013297 0.158516916 0.162367772 0.215862798 
#>                                                                         
#> 0.320793531 0.299984673 0.187336284 0.074116252 0.155882307 0.131519441 
#>                                                                         
#> 0.162270714 0.060187594 0.104681393 0.083020513 0.106465802 0.107090188 
#>                                                                         
#> 0.083786363 0.088353873 0.063246605 0.094278915 0.136746065 0.152764088 
#>                                                                         
#> 0.121214816 0.096714644 0.102026861 0.114521404 0.123418197 0.139155018 
#>                                                                         
#> 0.038258251 0.181375176 0.108609516 0.134408526 0.199741220 0.158947992 
#>                                                                         
#> 0.050630405 0.167705090 0.138038775 0.161164395 0.160660012 0.194250160 
#>                                                                         
#> 0.148415585 0.002481567 0.057894850 0.089442444 0.023040969 0.043976326 
#>                                                                         
#> 0.141666518 0.111720254 0.245009179 0.128593992 0.138410116 0.112533881 
#>                                                                         
#> 0.047592316 0.201417517 0.076426438 0.088244242 0.125418472 0.095440434 
#>                                                                         
#> 0.075214718 0.088807724 0.084584823 0.068202288 0.130823309 0.065634848 
#>                                                                         
#> 0.087020354 0.204547520 0.198627982 0.017507675 0.066559516 0.193886850 
#>                                                                         
#> 0.120515736 0.050528603 0.059167467 0.236492734 0.226930668 0.230948263 
#>                                                                         
#> 0.178816145 0.139744821 0.101922240 0.078968999 0.021606051 0.037438874 
#>                                                                         
#> 0.099256436 0.196436489 0.093857304 0.031049882 0.147124081 0.184683345 
#>                                                                         
#> 0.269680549 0.033408414 0.197845077 0.122208731 0.133320101 0.125213717 
#>                                                                         
#> 0.133069249 0.044021378 0.050065485 0.030635686 0.019730892 0.060965793 
#>                                                                         
#> 0.004768918 0.105631195 0.119337432 0.225492030 0.313434094 0.154494356 
#>                                                                         
#> 0.115645276 0.091827783 0.085284171 0.141789728 0.069195395 0.113163262 
#>                                                                         
#> 0.266929218 0.553018757 0.025068690 0.282130094 0.149971319 0.098742630 
#>                                                                         
#> 0.201543576 0.204697309 0.183587975 0.146538667 0.095555936 0.081155569 
#>                                                                         
#> 0.118271704 0.161058671 0.010644415 0.111582297 0.131756552 0.294276317 
#>                                                                         
#> 0.095945461 0.066092116 0.072458521 0.139346589 0.194547548 0.114352204 
#>                         
#> 0.514514448 0.189502225 
#> 
#> $alpha
#> [1] 0.2970766
#> 
#> attr(,"class")
#> [1] "scree"