Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1          y1         x2          y2          mx1
#>    34   38 0.30046045 0.929880273 0.30063886 0.925215737   0.32071571
#>     1   38 0.31474153 0.908902292 0.30063886 0.925215737   0.27580094
#>    85   90 0.36302476 0.845946590 0.41382078 0.849689156   0.38370847
#>     1   82 0.31474153 0.908902292 0.28139352 0.846341858   0.32229902
#>    85   82 0.36302476 0.845946590 0.28139352 0.846341858   0.32212372
#>    19   34 0.57543763 0.982514465 0.30046045 0.929880273   0.42915484
#>    85   88 0.36302476 0.845946590 0.36593658 0.820914463   0.38927035
#>    90   88 0.41382078 0.849689156 0.36593658 0.820914463   0.42723053
#>    82   71 0.28139352 0.846341858 0.16507463 0.711830965   0.26298212
#>    90   19 0.41382078 0.849689156 0.57543763 0.982514465   0.43189599
#>     5   19 0.62457887 0.786192377 0.57543763 0.982514465   0.53424945
#>    88   82 0.36593658 0.820914463 0.28139352 0.846341858   0.29634171
#>     3   71 0.02596811 0.817480595 0.16507463 0.711830965   0.15276393
#>    88   21 0.36593658 0.820914463 0.26953344 0.641640081   0.31391514
#>    62   88 0.43584549 0.695912674 0.36593658 0.820914463   0.34603726
#>    21   71 0.26953344 0.641640081 0.16507463 0.711830965   0.20832289
#>     3   50 0.02596811 0.817480595 0.01828184 0.785435902   0.10785485
#>    34    3 0.30046045 0.929880273 0.02596811 0.817480595   0.14352930
#>    29  100 0.08226385 0.685929167 0.14563194 0.680362380   0.11715869
#>     3   55 0.02596811 0.817480595 0.00000000 1.000000000  -0.26202076
#>    62    5 0.43584549 0.695912674 0.62457887 0.786192377   0.50924250
#>    62   94 0.43584549 0.695912674 0.39087331 0.599770795   0.43568079
#>    81   94 0.33325810 0.633316980 0.39087331 0.599770795   0.38736772
#>    81   21 0.33325810 0.633316980 0.26953344 0.641640081   0.29458048
#>    21  100 0.26953344 0.641640081 0.14563194 0.680362380   0.20743263
#>    50   30 0.01828184 0.785435902 0.04552775 0.715447384   0.10730242
#>    62   66 0.43584549 0.695912674 0.49398877 0.632513349   0.49658491
#>    21   35 0.26953344 0.641640081 0.14965235 0.630950138   0.21014164
#>    29   30 0.08226385 0.685929167 0.04552775 0.715447384  -0.01630176
#>    77   69 0.66158041 0.758222140 0.63477198 0.795134111   0.64436916
#>    94   16 0.39087331 0.599770795 0.35125812 0.561020550   0.44841726
#>    16   21 0.35125812 0.561020550 0.26953344 0.641640081   0.27639834
#>    35   29 0.14965235 0.630950138 0.08226385 0.685929167   0.06184467
#>    77   72 0.66158041 0.758222140 0.69241347 0.830217185   0.67406038
#>    76   77 0.54903100 0.662502428 0.66158041 0.758222140   0.60055356
#>    83   19 0.70814141 0.916661651 0.57543763 0.982514465   0.61285379
#>    16   89 0.35125812 0.561020550 0.20138839 0.570283253   0.27425653
#>    35   73 0.14965235 0.630950138 0.18665913 0.557208941   0.16742433
#>    72   83 0.69241347 0.830217185 0.70814141 0.916661651   0.61743554
#>    41   73 0.19535751 0.562101636 0.18665913 0.557208941   0.22810881
#>    32   15 0.56025797 0.611485591 0.68173418 0.654531427   0.61488609
#>    76   66 0.54903100 0.662502428 0.49398877 0.632513349   0.53017242
#>    66   94 0.49398877 0.632513349 0.39087331 0.599770795   0.47574026
#>    74   72 0.73852692 0.697464507 0.69241347 0.830217185   0.73339370
#>    15   74 0.68173418 0.654531427 0.73852692 0.697464507   0.68517864
#>    15   77 0.68173418 0.654531427 0.66158041 0.758222140   0.61762656
#>    25    9 0.42686469 0.384502027 0.46747298 0.388884373   0.43572369
#>    32   66 0.56025797 0.611485591 0.49398877 0.632513349   0.49143704
#>    89   41 0.20138839 0.570283253 0.19535751 0.562101636   0.26187131
#>    37   83 0.82856218 0.879761688 0.70814141 0.916661651   0.75756695
#>    75   41 0.31485283 0.460720515 0.19535751 0.562101636   0.23665094
#>    42   65 0.06096439 0.496102726 0.15446345 0.475580527   0.11889303
#>    35   42 0.14965235 0.630950138 0.06096439 0.496102726   0.10615926
#>    91   42 0.01078805 0.461385479 0.06096439 0.496102726  -0.04712946
#>     7   37 0.83201345 0.872179123 0.82856218 0.879761688   0.76385609
#>     7   72 0.83201345 0.872179123 0.69241347 0.830217185   0.77065258
#>    78   23 0.81164502 0.709522932 0.97917491 0.698030572   0.89764554
#>     4   74 0.80946328 0.754818804 0.73852692 0.697464507   0.77067341
#>    78   74 0.81164502 0.709522932 0.73852692 0.697464507   0.78055713
#>    75   25 0.31485283 0.460720515 0.42686469 0.384502027   0.41153260
#>    98   15 0.67435234 0.484367917 0.68173418 0.654531427   0.64297953
#>    75   65 0.31485283 0.460720515 0.15446345 0.475580527   0.22874550
#>    65   91 0.15446345 0.475580527 0.01078805 0.461385479   0.08997935
#>     9   51 0.46747298 0.388884373 0.59351869 0.418796255   0.50694806
#>    50   91 0.01828184 0.785435902 0.01078805 0.461385479  -0.27708539
#>    50   55 0.01828184 0.785435902 0.00000000 1.000000000  -2.47569851
#>    19   37 0.57543763 0.982514465 0.82856218 0.879761688   1.02758115
#>    67    7 0.88140499 0.869140776 0.83201345 0.872179123   0.85281380
#>    78   79 0.81164502 0.709522932 0.81334524 0.592423847   0.83529814
#>    43   15 0.77923545 0.620301590 0.68173418 0.654531427   0.70619491
#>    25   24 0.42686469 0.384502027 0.31685213 0.351124169   0.38907498
#>    51   98 0.59351869 0.418796255 0.67435234 0.484367917   0.58170291
#>    95   51 0.46782967 0.335934540 0.59351869 0.418796255   0.54021179
#>    98   57 0.67435234 0.484367917 0.64266114 0.434830076   0.68321151
#>    43   79 0.77923545 0.620301590 0.81334524 0.592423847   0.83299365
#>    98   36 0.67435234 0.484367917 0.71178345 0.473933055   0.71582847
#>    20   79 0.76371768 0.480744234 0.81334524 0.592423847   0.75238754
#>    36   20 0.71178345 0.473933055 0.76371768 0.480744234   0.72751130
#>    36   57 0.71178345 0.473933055 0.64266114 0.434830076   0.68604954
#>    57   56 0.64266114 0.434830076 0.61879079 0.362534590   0.64181114
#>    57   51 0.64266114 0.434830076 0.59351869 0.418796255   0.62685076
#>    86   20 0.77823925 0.441703004 0.76371768 0.480744234   0.74131128
#>    93   86 0.68624602 0.395210492 0.77823925 0.441703004   0.72901706
#>    23   67 0.97917491 0.698030572 0.88140499 0.869140776   0.90946433
#>    79   47 0.81334524 0.592423847 0.86399841 0.536604477   0.85390432
#>    93   57 0.68624602 0.395210492 0.64266114 0.434830076   0.64612649
#>    56   95 0.61879079 0.362534590 0.46782967 0.335934540   0.54397074
#>     6   31 0.85065300 0.429229916 0.87028886 0.522120409   0.83551057
#>    17   20 0.78161831 0.480761173 0.76371768 0.480744234   0.77268588
#>    26   86 0.68739358 0.359615224 0.77823925 0.441703004   0.75192504
#>    79   17 0.81334524 0.592423847 0.78161831 0.480761173   0.80541760
#>     2   14 0.89058295 0.621430617 0.90324809 0.596962971   0.92231323
#>     2   79 0.89058295 0.621430617 0.81334524 0.592423847   0.85856664
#>    22   26 0.63964999 0.355264058 0.68739358 0.359615224   0.66177509
#>    22   92 0.63964999 0.355264058 0.57501551 0.275308282   0.62552812
#>    68    8 0.01192768 0.325403452 0.01842530 0.298690194   0.09835186
#>    65   68 0.15446345 0.475580527 0.01192768 0.325403452   0.12412928
#>    91   55 0.01078805 0.461385479 0.00000000 1.000000000 -28.95527353
#>    14   79 0.90324809 0.596962971 0.81334524 0.592423847   0.85868991
#>    31   64 0.87028886 0.522120409 0.88716857 0.551189638   0.88095652
#>    47   17 0.86399841 0.536604477 0.78161831 0.480761173   0.82206223
#>    86   61 0.77823925 0.441703004 0.69175686 0.323351742   0.78592584
#>    78    2 0.81164502 0.709522932 0.89058295 0.621430617   0.89551177
#>    44   48 0.93193492 0.592093286 0.98556007 0.672730297   0.93884591
#>    14   64 0.90324809 0.596962971 0.88716857 0.551189638   0.91305548
#>    31   60 0.87028886 0.522120409 1.00000000 0.506502616   0.93397230
#>    24   28 0.31685213 0.351124169 0.44128145 0.179103512   0.38487426
#>    22   56 0.63964999 0.355264058 0.61879079 0.362534590   0.61254912
#>    24   65 0.31685213 0.351124169 0.15446345 0.475580527   0.18544934
#>    26   61 0.68739358 0.359615224 0.69175686 0.323351742   0.77680362
#>    48   23 0.98556007 0.672730297 0.97917491 0.698030572   0.92518014
#>    49   39 0.16208952 0.160069758 0.19451151 0.147899315   0.21925733
#>    60   44 1.00000000 0.506502616 0.93193492 0.592093286   0.94511742
#>     6   97 0.85065300 0.429229916 0.93471644 0.413599990   0.90117358
#>    86    6 0.77823925 0.441703004 0.85065300 0.429229916   0.81831221
#>    99   46 0.60915211 0.246523023 0.66744234 0.225799665   0.65420462
#>    61   22 0.69175686 0.323351742 0.63964999 0.355264058   0.63957121
#>    95   28 0.46782967 0.335934540 0.44128145 0.179103512   0.48946404
#>    92   95 0.57501551 0.275308282 0.46782967 0.335934540   0.49104291
#>    28   13 0.44128145 0.179103512 0.37437480 0.149041083   0.42122025
#>    99   92 0.60915211 0.246523023 0.57501551 0.275308282   0.55049597
#>    58   46 0.50038017 0.165241034 0.66744234 0.225799665   0.60407377
#>    58   28 0.50038017 0.165241034 0.44128145 0.179103512   0.47561385
#>    39   24 0.19451151 0.147899315 0.31685213 0.351124169   0.22467475
#>    87   49 0.07709532 0.142248942 0.16208952 0.160069758   0.11175090
#>    13   24 0.37437480 0.149041083 0.31685213 0.351124169   0.28390962
#>    60   48 1.00000000 0.506502616 0.98556007 0.672730297   1.01959661
#>    97   60 0.93471644 0.413599990 1.00000000 0.506502616   0.93163387
#>    40   97 0.92782473 0.361492848 0.93471644 0.413599990   0.88746478
#>    40   54 0.92782473 0.361492848 0.96096972 0.418368127   0.95354295
#>     8   87 0.01842530 0.298690194 0.07709532 0.142248942   0.02198299
#>    46   12 0.66744234 0.225799665 0.87165455 0.300584058   0.77399571
#>    12   40 0.87165455 0.300584058 0.92782473 0.361492848   0.86273414
#>    70   58 0.50469868 0.045319245 0.50038017 0.165241034   0.49482561
#>    63   70 0.40552922 0.075675992 0.50469868 0.045319245   0.46552479
#>    18   13 0.46456147 0.157382059 0.37437480 0.149041083   0.42198827
#>    39   80 0.19451151 0.147899315 0.23679720 0.035716777   0.28463778
#>    80   63 0.23679720 0.035716777 0.40552922 0.075675992   0.31525627
#>    52   12 0.86172433 0.152349730 0.87165455 0.300584058   0.78089267
#>    59   70 0.54217750 0.000000000 0.50469868 0.045319245   0.41644620
#>    84   52 0.76144185 0.095785154 0.86172433 0.152349730   0.76872955
#>    52   33 0.86172433 0.152349730 0.85011992 0.058007712   1.65911771
#>    52   40 0.86172433 0.152349730 0.92782473 0.361492848   1.02460417
#>    59   45 0.54217750 0.000000000 0.69755079 0.029095753   0.60561820
#>    80   59 0.23679720 0.035716777 0.54217750 0.000000000   0.36986200
#>    80   10 0.23679720 0.035716777 0.15110169 0.041638648  -0.09192928
#>    96   11 0.77129750 0.076610667 0.79747871 0.088860460   0.77958899
#>    45   96 0.69755079 0.029095753 0.77129750 0.076610667   0.72623239
#>    11   52 0.79747871 0.088860460 0.86172433 0.152349730   0.79217273
#>    45   27 0.69755079 0.029095753 0.74219856 0.002388074   0.73839110
#>    84   46 0.76144185 0.095785154 0.66744234 0.225799665   0.66451477
#>    27   33 0.74219856 0.002388074 0.85011992 0.058007712   0.80013114
#>    33   11 0.85011992 0.058007712 0.79747871 0.088860460   0.80430367
#>    33   60 0.85011992 0.058007712 1.00000000 0.506502616   2.80907037
#>    82   38 0.28139352 0.846341858 0.30063886 0.925215737   0.27580094
#>    82   34 0.28139352 0.846341858 0.30046045 0.929880273   0.14432162
#>     1   90 0.31474153 0.908902292 0.41382078 0.849689156   0.38370847
#>     1   34 0.31474153 0.908902292 0.30046045 0.929880273   0.42915484
#>     1   19 0.31474153 0.908902292 0.57543763 0.982514465   0.43189599
#>    21   82 0.26953344 0.641640081 0.28139352 0.846341858   0.29634171
#>    81   88 0.33325810 0.633316980 0.36593658 0.820914463   0.34603726
#>   100   71 0.14563194 0.680362380 0.16507463 0.711830965   0.20832289
#>    71   50 0.16507463 0.711830965 0.01828184 0.785435902   0.10785485
#>    82    3 0.28139352 0.846341858 0.02596811 0.817480595   0.14352930
#>    55   34 0.00000000 1.000000000 0.30046045 0.929880273   0.14137775
#>    55   19 0.00000000 1.000000000 0.57543763 0.982514465   0.30738095
#>    90    5 0.41382078 0.849689156 0.62457887 0.786192377   0.50924250
#>    35  100 0.14965235 0.630950138 0.14563194 0.680362380   0.20743263
#>    71   30 0.16507463 0.711830965 0.04552775 0.715447384   0.10730242
#>    69    5 0.63477198 0.795134111 0.62457887 0.786192377   0.55694345
#>    76   62 0.54903100 0.662502428 0.43584549 0.695912674   0.51814206
#>    89   35 0.20138839 0.570283253 0.14965235 0.630950138   0.21014164
#>    42   30 0.06096439 0.496102726 0.04552775 0.715447384  -0.01630176
#>    71   29 0.16507463 0.711830965 0.08226385 0.685929167   0.10651390
#>     9   16 0.46747298 0.388884373 0.35125812 0.561020550   0.44841726
#>    16   81 0.35125812 0.561020550 0.33325810 0.633316980   0.35223024
#>    89   21 0.20138839 0.570283253 0.26953344 0.641640081   0.27639834
#>    42   29 0.06096439 0.496102726 0.08226385 0.685929167   0.06184467
#>    69   72 0.63477198 0.795134111 0.69241347 0.830217185   0.67406038
#>     5   77 0.62457887 0.786192377 0.66158041 0.758222140   0.60055356
#>    69   19 0.63477198 0.795134111 0.57543763 0.982514465   0.61285379
#>    75   89 0.31485283 0.460720515 0.20138839 0.570283253   0.27425653
#>    89   73 0.20138839 0.570283253 0.18665913 0.557208941   0.16742433
#>    69   83 0.63477198 0.795134111 0.70814141 0.916661651   0.61743554
#>    65   73 0.15446345 0.475580527 0.18665913 0.557208941   0.22810881
#>    76   15 0.54903100 0.662502428 0.68173418 0.654531427   0.61488609
#>     9   94 0.46747298 0.388884373 0.39087331 0.599770795   0.47574026
#>    77   74 0.66158041 0.758222140 0.73852692 0.697464507   0.68517864
#>     9   66 0.46747298 0.388884373 0.49398877 0.632513349   0.49143704
#>    65   41 0.15446345 0.475580527 0.19535751 0.562101636   0.23665094
#>    73   42 0.18665913 0.557208941 0.06096439 0.496102726   0.10615926
#>    72   37 0.69241347 0.830217185 0.82856218 0.879761688   0.76385609
#>     4   72 0.80946328 0.754818804 0.69241347 0.830217185   0.77065258
#>     4   23 0.80946328 0.754818804 0.97917491 0.698030572   0.89764554
#>    43   74 0.77923545 0.620301590 0.73852692 0.697464507   0.78055713
#>    16   25 0.35125812 0.561020550 0.42686469 0.384502027   0.41153260
#>    68   91 0.01192768 0.325403452 0.01078805 0.461385479   0.08997935
#>    32   51 0.56025797 0.611485591 0.59351869 0.418796255   0.50694806
#>    30   91 0.04552775 0.715447384 0.01078805 0.461385479  -0.27708539
#>    67   37 0.88140499 0.869140776 0.82856218 0.879761688   1.02758115
#>     4    7 0.80946328 0.754818804 0.83201345 0.872179123   0.85281380
#>    95   24 0.46782967 0.335934540 0.31685213 0.351124169   0.38907498
#>    24   75 0.31685213 0.351124169 0.31485283 0.460720515   0.36005142
#>    32   98 0.56025797 0.611485591 0.67435234 0.484367917   0.58170291
#>    43   36 0.77923545 0.620301590 0.71178345 0.473933055   0.71582847
#>    95   25 0.46782967 0.335934540 0.42686469 0.384502027   0.44980245
#>    43   20 0.77923545 0.620301590 0.76371768 0.480744234   0.72751130
#>    56   51 0.61879079 0.362534590 0.59351869 0.418796255   0.62685076
#>    36   86 0.71178345 0.473933055 0.77823925 0.441703004   0.72901706
#>     4   67 0.80946328 0.754818804 0.88140499 0.869140776   0.90946433
#>    64   47 0.88716857 0.551189638 0.86399841 0.536604477   0.85390432
#>    22   57 0.63964999 0.355264058 0.64266114 0.434830076   0.64612649
#>    17   31 0.78161831 0.480761173 0.87028886 0.522120409   0.83551057
#>    44   14 0.93193492 0.592093286 0.90324809 0.596962971   0.92231323
#>    93   26 0.68624602 0.395210492 0.68739358 0.359615224   0.66177509
#>    53    8 0.10272759 0.245691369 0.01842530 0.298690194   0.09835186
#>    53   68 0.10272759 0.245691369 0.01192768 0.325403452   0.12412928
#>    68   55 0.01192768 0.325403452 0.00000000 1.000000000 -28.95527353
#>    64   79 0.88716857 0.551189638 0.81334524 0.592423847   0.85868991
#>    12   61 0.87165455 0.300584058 0.69175686 0.323351742   0.78592584
#>    23    2 0.97917491 0.698030572 0.89058295 0.621430617   0.89551177
#>     2   48 0.89058295 0.621430617 0.98556007 0.672730297   0.93884591
#>    44   64 0.93193492 0.592093286 0.88716857 0.551189638   0.91305548
#>    64   60 0.88716857 0.551189638 1.00000000 0.506502616   0.93397230
#>    92   56 0.57501551 0.275308282 0.61879079 0.362534590   0.61254912
#>    53   65 0.10272759 0.245691369 0.15446345 0.475580527   0.18544934
#>    53   39 0.10272759 0.245691369 0.19451151 0.147899315   0.21925733
#>    31   97 0.87028886 0.522120409 0.93471644 0.413599990   0.90117358
#>    17    6 0.78161831 0.480761173 0.85065300 0.429229916   0.81831221
#>    61   46 0.69175686 0.323351742 0.66744234 0.225799665   0.65420462
#>    99   22 0.60915211 0.246523023 0.63964999 0.355264058   0.63957121
#>    58   95 0.50038017 0.165241034 0.46782967 0.335934540   0.49104291
#>    12   86 0.87165455 0.300584058 0.77823925 0.441703004   0.80058939
#>    58   92 0.50038017 0.165241034 0.57501551 0.275308282   0.55049597
#>    18   28 0.46456147 0.157382059 0.44128145 0.179103512   0.47561385
#>    53   24 0.10272759 0.245691369 0.31685213 0.351124169   0.22467475
#>    53   49 0.10272759 0.245691369 0.16208952 0.160069758   0.11175090
#>    54   97 0.96096972 0.418368127 0.93471644 0.413599990   0.93581364
#>    10   87 0.15110169 0.041638648 0.07709532 0.142248942   0.12961510
#>    10   49 0.15110169 0.041638648 0.16208952 0.160069758   0.15829739
#>    53   87 0.10272759 0.245691369 0.07709532 0.142248942   0.02198299
#>     6   40 0.85065300 0.429229916 0.92782473 0.361492848   0.86273414
#>    45   58 0.69755079 0.029095753 0.50038017 0.165241034   0.61247229
#>    18   58 0.46456147 0.157382059 0.50038017 0.165241034   0.49482561
#>    18   70 0.46456147 0.157382059 0.50469868 0.045319245   0.46552479
#>    63   13 0.40552922 0.075675992 0.37437480 0.149041083   0.42198827
#>    13   80 0.37437480 0.149041083 0.23679720 0.035716777   0.28463778
#>    10   39 0.15110169 0.041638648 0.19451151 0.147899315   0.19713865
#>    46   52 0.66744234 0.225799665 0.86172433 0.152349730   0.76872955
#>    45   70 0.69755079 0.029095753 0.50469868 0.045319245   0.60716824
#>    54   33 0.96096972 0.418368127 0.85011992 0.058007712   1.65911771
#>    63   59 0.40552922 0.075675992 0.54217750 0.000000000   0.36986200
#>    59   10 0.54217750 0.000000000 0.15110169 0.041638648  -0.09192928
#>    84   11 0.76144185 0.095785154 0.79747871 0.088860460   0.77958899
#>    84   96 0.76144185 0.095785154 0.77129750 0.076610667   0.72623239
#>    96   27 0.77129750 0.076610667 0.74219856 0.002388074   0.73839110
#>    45   46 0.69755079 0.029095753 0.66744234 0.225799665   0.66451477
#>    96   33 0.77129750 0.076610667 0.85011992 0.058007712   0.80013114
#>    27   59 0.74219856 0.002388074 0.54217750 0.000000000   0.64353794
#>    54   60 0.96096972 0.418368127 1.00000000 0.506502616   2.80907037
#>     1   85 0.31474153 0.908902292 0.36302476 0.845946590   0.38370847
#>    62   90 0.43584549 0.695912674 0.41382078 0.849689156   0.42723053
#>    81   62 0.33325810 0.633316980 0.43584549 0.695912674   0.34603726
#>    76    5 0.54903100 0.662502428 0.62457887 0.786192377   0.51814206
#>    75   16 0.31485283 0.460720515 0.35125812 0.561020550   0.27425653
#>    76   32 0.54903100 0.662502428 0.56025797 0.611485591   0.61488609
#>     9   32 0.46747298 0.388884373 0.56025797 0.611485591   0.49143704
#>     4   78 0.80946328 0.754818804 0.81164502 0.709522932   0.89764554
#>    43   78 0.77923545 0.620301590 0.81164502 0.709522932   0.78055713
#>    67   19 0.88140499 0.869140776 0.57543763 0.982514465   1.02758115
#>    98   43 0.67435234 0.484367917 0.77923545 0.620301590   0.70619491
#>     9   95 0.46747298 0.388884373 0.46782967 0.335934540   0.54021179
#>    93   36 0.68624602 0.395210492 0.71178345 0.473933055   0.68604954
#>    22   93 0.63964999 0.355264058 0.68624602 0.395210492   0.64612649
#>    86   17 0.77823925 0.441703004 0.78161831 0.480761173   0.77268588
#>    44    2 0.93193492 0.592093286 0.89058295 0.621430617   0.92231323
#>    47   31 0.86399841 0.536604477 0.87028886 0.522120409   0.88095652
#>    61   99 0.69175686 0.323351742 0.60915211 0.246523023   0.65420462
#>    12    6 0.87165455 0.300584058 0.85065300 0.429229916   0.80058939
#>    58   99 0.50038017 0.165241034 0.60915211 0.246523023   0.55049597
#>    39   13 0.19451151 0.147899315 0.37437480 0.149041083   0.28390962
#>    18   63 0.46456147 0.157382059 0.40552922 0.075675992   0.46552479
#>    54   52 0.96096972 0.418368127 0.86172433 0.152349730   1.65911771
#>    84   45 0.76144185 0.095785154 0.69755079 0.029095753   0.72623239
#>           my1          mx2         my2 bp1 bp2
#>   0.928319332   0.14432162  0.92157247   0   0
#>   0.889491340   0.32071571  0.92831933   0   0
#>   0.911802764   0.38927035  0.83631415   0   0
#>   0.864705415   0.27580094  0.88949134   0   0
#>   0.828503421   0.32229902  0.86470542   0   0
#>   1.002140951   0.30738095  1.63832520   0   0
#>   0.836314147   0.32212372  0.82850342   0   0
#>   0.773144283   0.38927035  0.83631415   0   0
#>   0.744714119   0.15276393  0.84002578   0   0
#>   0.992433238   0.53424945  0.86789338   0   0
#>   0.867893384   0.55694345  0.87357390   0   0
#>   0.742781319   0.32212372  0.82850342   0   0
#>   0.840025784   0.10785485  0.78089499   0   0
#>   0.733331370   0.29634171  0.74278132   0   0
#>   0.727735868   0.42723053  0.77314428   0   0
#>   0.663369693   0.26298212  0.74471412   0   0
#>   0.780894993  -0.26202076  0.86961374   0   0
#>   0.921753315   0.14137775  0.92700763   0   0
#>   0.719694995   0.11127303  0.65269710   0   0
#>   0.869613739   0.14137775  0.92700763   0   0
#>   0.784890484   0.51814206  0.76628561   0   0
#>   0.637400481   0.38736772  0.65999983   0   0
#>   0.659999832   0.35223024  0.59965158   0   0
#>   0.585298216   0.31391514  0.73333137   0   0
#>   0.660521090   0.20832289  0.66336969   0   0
#>   0.779793267  -0.27708539  0.63015452   0   0
#>   0.693255415   0.43568079  0.63740048   0   0
#>   0.630141322   0.20743263  0.66052109   0   0
#>   0.600880545   0.10651390  0.75372750   0   0
#>   0.773913153   0.67406038  0.79547729   0   0
#>   0.501317568   0.35223024  0.59965158   0   0
#>   0.566866835   0.29458048  0.58529822   0   0
#>   0.592112127   0.11127303  0.65269710   0   0
#>   0.795477291   0.73339370  0.77006682   0   0
#>   0.715949957   0.61762656  0.69587514   0   0
#>   0.891278019   1.01382465  1.69929732   0   0
#>   0.532212428   0.27639834  0.56686683   0   0
#>   0.593712481   0.10615926  0.56296681   0   0
#>   0.888511896   0.75756695  0.86301601   0   0
#>   0.493696980   0.17951050  0.58009644   0   0
#>   0.650251019   0.64297953  0.57097077   0   0
#>   0.631608588   0.49658491  0.69325541   0   0
#>   0.511242024   0.43568079  0.63740048   0   0
#>   0.770066821   0.73736489  0.77144627   0   0
#>   0.709004851   0.73327929  0.64537636   0   0
#>   0.695875137   0.68517864  0.70900485   0   0
#>   0.492747721   0.44980245  0.36228922   0   0
#>   0.509533637   0.53017242  0.63160859   0   0
#>   0.519386159   0.17951050  0.58009644   0   0
#>   0.863016010   1.01382465  1.69929732   0   0
#>   0.489659553   0.26187131  0.51938616   0   0
#>   0.536773608   0.08923061  0.40163177   0   0
#>   0.562966806   0.06184467  0.59211213   0   0
#>   0.598711009   0.08923061  0.40163177   0   0
#>   0.845733398   0.85780640  0.88849574   0   0
#>   0.823122683   0.76385609  0.84573340   0   0
#>   0.736365748   0.89551177  0.70526083   0   0
#>   0.730249955   0.73736489  0.77144627   0   0
#>   0.670318538   0.77067341  0.73024996   0   0
#>   0.482386170   0.36005142  0.40672864   0   0
#>   0.570970767   0.70619491  0.56822843   0   0
#>   0.404333303   0.23665094  0.48965955   0   0
#>   0.394053373   0.08923061  0.40163177   0   0
#>   0.503068311   0.54021179  0.36289825   0   0
#>   0.630154523  -2.47569851  0.68099830   0   0
#>   0.680998295  -0.26202076  0.86961374   0   0
#>   1.733185538   1.01382465  1.69929732   0   0
#>   0.807335837   0.85780640  0.88849574   0   0
#>   0.651304477   0.83299365  0.65127102   0   0
#>   0.568228430   0.73327929  0.64537636   0   0
#>   0.311067710   0.36005142  0.40672864   0   0
#>   0.515971982   0.59390953  0.50092424   0   0
#>   0.362898250   0.54115463  0.36146811   0   0
#>   0.443794436   0.59390953  0.50092424   0   0
#>   0.651271017   0.75238754  0.55264547   0   0
#>   0.560795408   0.68321151  0.44379444   0   0
#>   0.552645473   0.77260847  0.54365981   0   0
#>   0.555411530   0.74131128  0.45018877   0   0
#>   0.438777656   0.68321151  0.44379444   0   0
#>   0.395022257   0.62685076  0.39996184   0   0
#>   0.399961840   0.59390953  0.50092424   0   0
#>   0.450188774   0.77268588  0.46185870   0   0
#>   0.424839081   0.75192504  0.37951180   0   0
#>   0.771686212   5.31524487  3.28908614   0   1
#>   0.578336856   0.80541760  0.53433769   0   0
#>   0.394858943   0.68604954  0.43877766   0   0
#>   0.345486024   0.54115463  0.36146811   0   0
#>   0.480951463   0.90117358  0.46707114   0   0
#>   0.461858699   0.77260847  0.54365981   0   0
#>   0.379511803   0.77680362  0.35197894   0   0
#>   0.534337689   0.77260847  0.54365981   0   0
#>   0.622343361   0.85856664  0.58934634   0   0
#>   0.589346335   0.83529814  0.65130448   0   0
#>   0.376605435   0.66524279  0.33855577   0   0
#>   0.300577439   0.61254912  0.31106937   0   0
#>   0.332278036 -35.24519341 -8.26453154   0   1
#>   0.361641060   0.08997935  0.39405337   0   0
#>   0.150632037  -2.47569851  0.68099830   0   0
#>   0.586904705   0.85856664  0.58934634   0   0
#>   0.535361402   0.93397230  0.50457657   0   0
#>   0.509783512   0.80541760  0.53433769   0   0
#>   0.345313096   0.77680362  0.35197894   0   0
#>   0.705260826   0.83529814  0.65130448   0   0
#>   0.645646727   1.01959661  0.59194597   0   0
#>   0.567806856   0.85868991  0.58690471   0   0
#>   0.504576574   0.93163387  0.48515514   0   0
#>   0.269314618   0.36650949  0.25603065   0   0
#>   0.311069371   0.64181114  0.39502226   0   0
#>   0.347841166   0.22874550  0.40433330   0   0
#>   0.351978939   0.66524279  0.33855577   0   0
#>   0.670947768  38.08369640 10.04883411   0   1
#>   0.263093298   0.15829739  0.10069631   0   0
#>   0.532717177   1.01959661  0.59194597   0   0
#>   0.467071145   0.88746478  0.39334019   0   0
#>   0.457911395   0.80058939  0.35501980   0   0
#>   0.280905315   0.60407377  0.13989832   0   0
#>   0.296638866   0.66524279  0.33855577   0   0
#>   0.251609747   0.38487426  0.26931462   0   0
#>   0.251910831   0.54397074  0.34548602   0   0
#>   0.134266921   0.36650949  0.25603065   0   0
#>   0.211596394   0.62552812  0.30057744   0   0
#>   0.139898324   0.61247229  0.11672946   0   0
#>   0.192563423   0.48946404  0.25160975   0   0
#>   0.268177883   0.28390962  0.23251871   0   0
#>   0.188558502   0.12961510  0.10335740   0   0
#>   0.232518712   0.36650949  0.25603065   0   0
#>   0.591945968   6.99038558  1.11061864   0   1
#>   0.485155141   0.93581364  0.48221798   0   0
#>   0.393340191   0.95354295  0.38460067   0   0
#>   0.384600668   1.50189798  0.06503785   0   0
#>   0.210802324  -5.58200920 -1.89085782   0   1
#>   0.251047799   0.78089267  0.23221442   0   0
#>   0.365164983   1.02460417  0.21588819   0   0
#>   0.105002358   0.60716824  0.10904793   0   0
#>   0.094507769   0.41644620 -0.06582223   0   0
#>   0.125962705   0.42122025  0.13426692   0   0
#>   0.117810412   0.19713865  0.08482880   0   0
#>   0.080639060   0.36986200 -0.14993941   0   0
#>   0.232214418   1.02460417  0.21588819   0   0
#>  -0.065822227   0.60561820  0.09062223   0   0
#>   0.200041832   0.79217273  0.15847978   0   0
#>   0.006382761   0.84331365  0.10672961   0   0
#>   0.215888191   1.50189798  0.06503785   0   0
#>   0.090622228   0.64353794 -0.11187177   0   0
#>  -0.149939408  -0.09192928 -4.09827886   0   0
#>  -4.098278855   0.19713865  0.08482880   0   0
#>   0.092992608   0.80430367  0.04017046   0   0
#>   0.065567431   0.73839110  0.04669620   0   0
#>   0.158479776   0.84331365  0.10672961   0   0
#>   0.046696197   0.64353794 -0.11187177   0   0
#>   0.124695328   0.76872955  0.20004183   0   0
#>   0.022491042   4.57336782 -7.29889498   0   1
#>   0.040170459   0.84331365  0.10672961   0   0
#>  -0.347352042   4.81475644 -1.01762138   0   1
#>   0.889491340   0.14432162  0.92157247   0   0
#>   0.921572475   0.14352930  0.92175332   0   0
#>   0.911802764   0.43189599  0.99243324   0   0
#>   1.002140951   0.32071571  0.92831933   0   0
#>   0.992433238   0.42915484  1.00214095   0   0
#>   0.742781319   0.26298212  0.74471412   0   0
#>   0.727735868   0.31391514  0.73333137   0   0
#>   0.663369693   0.11715869  0.71969500   0   0
#>   0.780894993   0.10730242  0.77979327   0   0
#>   0.921753315   0.15276393  0.84002578   0   0
#>   0.927007626   0.30738095  1.63832520   0   0
#>   1.638325197   0.36013813  3.37452987   0   1
#>   0.784890484   0.53424945  0.86789338   0   0
#>   0.660521090   0.11127303  0.65269710   0   0
#>   0.779793267   0.10651390  0.75372750   0   0
#>   0.873573902   0.64436916  0.77391315   0   0
#>   0.766285606   0.49658491  0.69325541   0   0
#>   0.630141322   0.16742433  0.59371248   0   0
#>   0.600880545  -0.04712946  0.59871101   0   0
#>   0.753727497   0.11715869  0.71969500   0   0
#>   0.501317568   0.43572369  0.49274772   0   0
#>   0.599651578   0.29458048  0.58529822   0   0
#>   0.566866835   0.21014164  0.63014132   0   0
#>   0.592112127  -0.01630176  0.60088054   0   0
#>   0.795477291   0.61743554  0.88851190   0   0
#>   0.715949957   0.64436916  0.77391315   0   0
#>   0.891278019   0.55694345  0.87357390   0   0
#>   0.532212428   0.26187131  0.51938616   0   0
#>   0.593712481   0.17951050  0.58009644   0   0
#>   0.888511896   0.61285379  0.89127802   0   0
#>   0.493696980   0.11889303  0.53677361   0   0
#>   0.650251019   0.61762656  0.69587514   0   0
#>   0.511242024   0.44841726  0.50131757   0   0
#>   0.709004851   0.73339370  0.77006682   0   0
#>   0.509533637   0.47574026  0.51124202   0   0
#>   0.489659553   0.22810881  0.49369698   0   0
#>   0.562966806   0.11889303  0.53677361   0   0
#>   0.845733398   0.75756695  0.86301601   0   0
#>   0.823122683   0.73736489  0.77144627   0   0
#>   0.736365748   0.90946433  0.77168621   0   0
#>   0.670318538   0.73327929  0.64537636   0   0
#>   0.482386170   0.43572369  0.49274772   0   0
#>   0.394053373 -28.95527353  0.15063204   0   0
#>   0.503068311   0.58170291  0.51597198   0   0
#>   0.630154523  -0.04712946  0.59871101   0   0
#>   1.733185538   0.85780640  0.88849574   0   0
#>   0.807335837   0.77065258  0.82312268   0   0
#>   0.311067710   0.38487426  0.26931462   0   0
#>   0.406728636   0.22874550  0.40433330   0   0
#>   0.515971982   0.64297953  0.57097077   0   0
#>   0.560795408   0.72751130  0.55541153   0   0
#>   0.362289220   0.38907498  0.31106771   0   0
#>   0.555411530   0.75238754  0.55264547   0   0
#>   0.399961840   0.54115463  0.36146811   0   0
#>   0.424839081   0.74131128  0.45018877   0   0
#>   0.771686212   0.85281380  0.80733584   0   0
#>   0.578336856   0.88095652  0.53536140   0   0
#>   0.394858943   0.64181114  0.39502226   0   0
#>   0.480951463   0.82206223  0.50978351   0   0
#>   0.622343361   0.91305548  0.56780686   0   0
#>   0.376605435   0.75192504  0.37951180   0   0
#>   0.332278036   0.02198299  0.21080232   0   0
#>   0.361641060   0.09835186  0.33227804   0   0
#>   0.150632037 -30.43717775  0.12443018   0   1
#>   0.586904705   0.85390432  0.57833686   0   0
#>   0.345313096   0.77399571  0.25104780   0   0
#>   0.705260826   0.92518014  0.67094777   0   0
#>   0.645646727   0.92518014  0.67094777   0   0
#>   0.567806856   0.94511742  0.53271718   0   0
#>   0.504576574   0.94511742  0.53271718   0   0
#>   0.311069371   0.54397074  0.34548602   0   0
#>   0.347841166   0.12412928  0.36164106   0   0
#>   0.263093298   0.22467475  0.26817788   0   0
#>   0.467071145   0.93163387  0.48515514   0   0
#>   0.457911395   0.83551057  0.48095146   0   0
#>   0.280905315   0.77399571  0.25104780   0   0
#>   0.296638866   0.62552812  0.30057744   0   0
#>   0.251910831   0.48946404  0.25160975   0   0
#>   0.355019800   0.78592584  0.34531310   0   0
#>   0.211596394   0.49104291  0.25191083   0   0
#>   0.192563423   0.42122025  0.13426692   0   0
#>   0.268177883   0.18544934  0.34784117   0   0
#>   0.188558502   0.21925733  0.26309330   0   0
#>   0.482217981   0.95354295  0.38460067   0   0
#>   0.103357405  -6.32002129 -4.64083105   0   1
#>   0.100696314   0.12961510  0.10335740   0   0
#>   0.210802324   0.11175090  0.18855850   0   0
#>   0.365164983   0.88746478  0.39334019   0   0
#>   0.116729455   0.60716824  0.10904793   0   0
#>   0.105002358   0.47561385  0.19256342   0   0
#>   0.094507769   0.49482561  0.10500236   0   0
#>   0.125962705   0.31525627  0.08063906   0   0
#>   0.117810412   0.31525627  0.08063906   0   0
#>   0.084828803   0.15829739  0.10069631   0   0
#>   0.200041832   0.78089267  0.23221442   0   0
#>   0.109047928   0.60561820  0.09062223   0   0
#>   0.006382761   2.80907037 -0.34735204   0   0
#>  -0.149939408   0.41644620 -0.06582223   0   0
#>  -4.098278855  -0.36113179 -6.62666529   0   1
#>   0.092992608   0.79217273  0.15847978   0   0
#>   0.065567431   0.77958899  0.09299261   0   0
#>   0.046696197   0.80013114  0.02249104   0   0
#>   0.124695328   0.61247229  0.11672946   0   0
#>   0.022491042   0.80430367  0.04017046   0   0
#>  -0.111871769   0.70321872 -5.11063279   0   1
#>  -0.347352042   0.93581364  0.48221798   0   0
#>   0.911802764   0.32229902  0.86470542   0   0
#>   0.773144283   0.50924250  0.78489048   0   0
#>   0.727735868   0.38736772  0.65999983   0   0
#>   0.766285606   0.60055356  0.71594996   0   0
#>   0.532212428   0.41153260  0.48238617   0   0
#>   0.650251019   0.53017242  0.63160859   0   0
#>   0.509533637   0.50694806  0.50306831   0   0
#>   0.736365748   0.77067341  0.73024996   0   0
#>   0.670318538   0.83299365  0.65127102   0   0
#>   1.733185538   2.09242662  4.60693831   0   1
#>   0.568228430   0.71582847  0.56079541   0   0
#>   0.362898250   0.44980245  0.36228922   0   0
#>   0.438777656   0.72901706  0.42483908   0   0
#>   0.394858943   0.66177509  0.37660544   0   0
#>   0.461858699   0.81831221  0.45791140   0   0
#>   0.622343361   0.93884591  0.64564673   0   0
#>   0.535361402   0.82206223  0.50978351   0   0
#>   0.280905315   0.63957121  0.29663887   0   0
#>   0.355019800   0.86273414  0.36516498   0   0
#>   0.211596394   0.60407377  0.13989832   0   0
#>   0.232518712   0.28463778  0.11781041   0   0
#>   0.094507769   0.42198827  0.12596271   0   0
#>   0.006382761   1.50189798  0.06503785   0   0
#>   0.065567431   0.66451477  0.12469533   0   0
#> 
#> $x
#>              [,1]        [,2]
#>   [1,] 0.31474153 0.908902292
#>   [2,] 0.89058295 0.621430617
#>   [3,] 0.02596811 0.817480595
#>   [4,] 0.80946328 0.754818804
#>   [5,] 0.62457887 0.786192377
#>   [6,] 0.85065300 0.429229916
#>   [7,] 0.83201345 0.872179123
#>   [8,] 0.01842530 0.298690194
#>   [9,] 0.46747298 0.388884373
#>  [10,] 0.15110169 0.041638648
#>  [11,] 0.79747871 0.088860460
#>  [12,] 0.87165455 0.300584058
#>  [13,] 0.37437480 0.149041083
#>  [14,] 0.90324809 0.596962971
#>  [15,] 0.68173418 0.654531427
#>  [16,] 0.35125812 0.561020550
#>  [17,] 0.78161831 0.480761173
#>  [18,] 0.46456147 0.157382059
#>  [19,] 0.57543763 0.982514465
#>  [20,] 0.76371768 0.480744234
#>  [21,] 0.26953344 0.641640081
#>  [22,] 0.63964999 0.355264058
#>  [23,] 0.97917491 0.698030572
#>  [24,] 0.31685213 0.351124169
#>  [25,] 0.42686469 0.384502027
#>  [26,] 0.68739358 0.359615224
#>  [27,] 0.74219856 0.002388074
#>  [28,] 0.44128145 0.179103512
#>  [29,] 0.08226385 0.685929167
#>  [30,] 0.04552775 0.715447384
#>  [31,] 0.87028886 0.522120409
#>  [32,] 0.56025797 0.611485591
#>  [33,] 0.85011992 0.058007712
#>  [34,] 0.30046045 0.929880273
#>  [35,] 0.14965235 0.630950138
#>  [36,] 0.71178345 0.473933055
#>  [37,] 0.82856218 0.879761688
#>  [38,] 0.30063886 0.925215737
#>  [39,] 0.19451151 0.147899315
#>  [40,] 0.92782473 0.361492848
#>  [41,] 0.19535751 0.562101636
#>  [42,] 0.06096439 0.496102726
#>  [43,] 0.77923545 0.620301590
#>  [44,] 0.93193492 0.592093286
#>  [45,] 0.69755079 0.029095753
#>  [46,] 0.66744234 0.225799665
#>  [47,] 0.86399841 0.536604477
#>  [48,] 0.98556007 0.672730297
#>  [49,] 0.16208952 0.160069758
#>  [50,] 0.01828184 0.785435902
#>  [51,] 0.59351869 0.418796255
#>  [52,] 0.86172433 0.152349730
#>  [53,] 0.10272759 0.245691369
#>  [54,] 0.96096972 0.418368127
#>  [55,] 0.00000000 1.000000000
#>  [56,] 0.61879079 0.362534590
#>  [57,] 0.64266114 0.434830076
#>  [58,] 0.50038017 0.165241034
#>  [59,] 0.54217750 0.000000000
#>  [60,] 1.00000000 0.506502616
#>  [61,] 0.69175686 0.323351742
#>  [62,] 0.43584549 0.695912674
#>  [63,] 0.40552922 0.075675992
#>  [64,] 0.88716857 0.551189638
#>  [65,] 0.15446345 0.475580527
#>  [66,] 0.49398877 0.632513349
#>  [67,] 0.88140499 0.869140776
#>  [68,] 0.01192768 0.325403452
#>  [69,] 0.63477198 0.795134111
#>  [70,] 0.50469868 0.045319245
#>  [71,] 0.16507463 0.711830965
#>  [72,] 0.69241347 0.830217185
#>  [73,] 0.18665913 0.557208941
#>  [74,] 0.73852692 0.697464507
#>  [75,] 0.31485283 0.460720515
#>  [76,] 0.54903100 0.662502428
#>  [77,] 0.66158041 0.758222140
#>  [78,] 0.81164502 0.709522932
#>  [79,] 0.81334524 0.592423847
#>  [80,] 0.23679720 0.035716777
#>  [81,] 0.33325810 0.633316980
#>  [82,] 0.28139352 0.846341858
#>  [83,] 0.70814141 0.916661651
#>  [84,] 0.76144185 0.095785154
#>  [85,] 0.36302476 0.845946590
#>  [86,] 0.77823925 0.441703004
#>  [87,] 0.07709532 0.142248942
#>  [88,] 0.36593658 0.820914463
#>  [89,] 0.20138839 0.570283253
#>  [90,] 0.41382078 0.849689156
#>  [91,] 0.01078805 0.461385479
#>  [92,] 0.57501551 0.275308282
#>  [93,] 0.68624602 0.395210492
#>  [94,] 0.39087331 0.599770795
#>  [95,] 0.46782967 0.335934540
#>  [96,] 0.77129750 0.076610667
#>  [97,] 0.93471644 0.413599990
#>  [98,] 0.67435234 0.484367917
#>  [99,] 0.60915211 0.246523023
#> [100,] 0.14563194 0.680362380
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,34,38), [3,2,7]
#> 2: (82,1,38), [1,3,5]
#> 3: (82,38,34), [1,19,2]
#> 4: (1,85,90), [8,11,5]
#> 5: (85,1,82), [2,6,4]
#> 6: (88,85,82), [5,13,8]
#> 7: (1,19,34), [21,1,11]
#> 8: (90,85,88), [6,9,4]
#> 9: (62,90,88), [8,16,24]
#> 10: (21,82,71), [14,17,13]
#> 11: (1,90,19), [12,7,4]
#> 12: (90,5,19), [31,11,24]
#> 13: (21,88,82), [6,10,15]
#> 14: (82,3,71), [18,10,19]
#> 15: (81,88,21), [13,27,16]
#> 16: (81,62,88), [9,15,26]
#> 17: (100,21,71), [10,22,29]
#> 18: (71,3,50), [23,30,14]
#> 19: (82,34,3), [20,14,3]
#> 20: (55,3,34), [19,21,23]
#> 21: (55,34,19), [7,0,20]
#> 22: (71,29,100), [28,17,36]
#> 23: (50,3,55), [20,79,18]
#> 24: (90,62,5), [32,12,9]
#> 25: (66,62,94), [26,52,33]
#> 26: (62,81,94), [39,25,16]
#> 27: (16,81,21), [15,40,39]
#> 28: (35,100,29), [22,41,29]
#> 29: (35,21,100), [17,28,34]
#> 30: (71,50,30), [78,36,18]
#> 31: (69,19,5), [12,37,44]
#> 32: (76,5,62), [24,33,43]
#> 33: (76,62,66), [25,51,32]
#> 34: (89,21,35), [29,46,40]
#> 35: (42,29,30), [36,63,41]
#> 36: (71,30,29), [35,22,30]
#> 37: (5,77,69), [42,31,43]
#> 38: (9,94,16), [39,56,52]
#> 39: (16,94,81), [26,27,38]
#> 40: (89,16,21), [27,34,45]
#> 41: (42,35,29), [28,35,62]
#> 42: (69,77,72), [53,47,37]
#> 43: (5,76,77), [55,37,32]
#> 44: (69,83,19), [64,31,47]
#> 45: (75,16,89), [40,58,72]
#> 46: (89,35,73), [62,49,34]
#> 47: (69,72,83), [59,44,42]
#> 48: (65,41,73), [49,61,60]
#> 49: (89,73,41), [48,58,46]
#> 50: (76,32,15), [73,55,51]
#> 51: (32,76,66), [33,57,50]
#> 52: (9,66,94), [25,38,57]
#> 53: (77,74,72), [65,42,54]
#> 54: (77,15,74), [71,53,55]
#> 55: (76,15,77), [54,43,50]
#> 56: (16,25,9), [95,38,72]
#> 57: (9,32,66), [51,52,76]
#> 58: (75,89,41), [49,60,45]
#> 59: (72,37,83), [64,47,66]
#> 60: (65,75,41), [58,48,74]
#> 61: (73,42,65), [77,48,62]
#> 62: (73,35,42), [41,61,46]
#> 63: (30,91,42), [77,35,78]
#> 64: (37,19,83), [44,59,80]
#> 65: (4,72,74), [53,69,67]
#> 66: (72,7,37), [81,59,67]
#> 67: (4,7,72), [66,65,82]
#> 68: (4,78,23), [122,103,69]
#> 69: (78,4,74), [65,70,68]
#> 70: (43,78,74), [69,71,91]
#> 71: (43,74,15), [54,84,70]
#> 72: (16,75,25), [86,56,45]
#> 73: (32,98,15), [84,50,87]
#> 74: (24,75,65), [60,128,86]
#> 75: (68,65,91), [77,117,116]
#> 76: (32,9,51), [88,87,57]
#> 77: (65,42,91), [63,75,61]
#> 78: (30,50,91), [79,63,30]
#> 79: (91,50,55), [23,117,78]
#> 80: (67,19,37), [64,81,0]
#> 81: (67,37,7), [66,82,80]
#> 82: (4,67,7), [81,67,103]
#> 83: (2,78,79), [91,112,122]
#> 84: (98,43,15), [71,73,92]
#> 85: (95,25,24), [86,126,95]
#> 86: (24,25,75), [72,74,85]
#> 87: (32,51,98), [90,73,76]
#> 88: (9,95,51), [94,76,95]
#> 89: (36,98,57), [90,97,92]
#> 90: (57,98,51), [87,99,89]
#> 91: (78,43,79), [93,83,70]
#> 92: (43,98,36), [89,96,84]
#> 93: (43,20,79), [100,91,96]
#> 94: (56,51,95), [88,106,99]
#> 95: (95,9,25), [56,85,88]
#> 96: (43,36,20), [101,93,92]
#> 97: (93,36,57), [89,105,102]
#> 98: (22,57,56), [99,127,105]
#> 99: (56,57,51), [90,94,98]
#> 100: (17,79,20), [93,108,110]
#> 101: (36,86,20), [108,96,102]
#> 102: (36,93,86), [109,101,97]
#> 103: (4,23,67), [0,82,68]
#> 104: (64,79,47), [110,119,118]
#> 105: (22,93,57), [97,98,113]
#> 106: (92,56,95), [94,139,127]
#> 107: (17,6,31), [134,120,135]
#> 108: (86,17,20), [100,101,135]
#> 109: (93,26,86), [129,102,113]
#> 110: (47,79,17), [100,120,104]
#> 111: (44,2,14), [112,124,123]
#> 112: (14,2,79), [83,118,111]
#> 113: (93,22,26), [130,109,105]
#> 114: (99,22,92), [127,143,137]
#> 115: (53,68,8), [0,156,116]
#> 116: (53,65,68), [75,115,128]
#> 117: (68,91,55), [79,0,75]
#> 118: (64,14,79), [112,104,124]
#> 119: (47,31,64), [125,104,120]
#> 120: (31,47,17), [110,107,119]
#> 121: (12,86,61), [129,157,142]
#> 122: (23,78,2), [83,131,68]
#> 123: (2,44,48), [149,131,111]
#> 124: (44,14,64), [118,133,111]
#> 125: (64,31,60), [150,133,119]
#> 126: (95,24,28), [141,138,85]
#> 127: (92,22,56), [98,106,114]
#> 128: (53,24,65), [74,116,146]
#> 129: (86,26,61), [130,121,109]
#> 130: (61,26,22), [113,137,129]
#> 131: (2,48,23), [0,122,123]
#> 132: (53,49,39), [155,146,147]
#> 133: (64,60,44), [149,124,125]
#> 134: (31,6,97), [151,150,107]
#> 135: (17,86,6), [142,107,108]
#> 136: (61,99,46), [144,157,137]
#> 137: (99,61,22), [130,114,136]
#> 138: (58,95,28), [126,145,139]
#> 139: (58,92,95), [106,138,143]
#> 140: (18,28,13), [141,162,145]
#> 141: (13,28,24), [126,148,140]
#> 142: (12,6,86), [135,121,158]
#> 143: (58,99,92), [114,139,144]
#> 144: (99,58,46), [159,136,143]
#> 145: (18,58,28), [138,140,160]
#> 146: (53,39,24), [148,128,132]
#> 147: (53,87,49), [154,132,156]
#> 148: (39,13,24), [141,146,163]
#> 149: (44,60,48), [0,123,133]
#> 150: (31,97,60), [152,125,134]
#> 151: (6,40,97), [153,134,158]
#> 152: (54,60,97), [150,153,184]
#> 153: (97,40,54), [171,152,151]
#> 154: (10,49,87), [147,0,155]
#> 155: (10,39,49), [132,154,167]
#> 156: (53,8,87), [0,147,115]
#> 157: (61,46,12), [165,121,136]
#> 158: (6,12,40), [172,151,142]
#> 159: (45,46,58), [144,169,180]
#> 160: (18,70,58), [169,145,161]
#> 161: (18,63,70), [166,160,162]
#> 162: (63,18,13), [140,164,161]
#> 163: (13,39,80), [167,164,148]
#> 164: (13,80,63), [174,162,163]
#> 165: (46,52,12), [172,157,168]
#> 166: (63,59,70), [173,161,174]
#> 167: (10,80,39), [163,155,175]
#> 168: (46,84,52), [178,165,180]
#> 169: (45,58,70), [160,173,159]
#> 170: (54,52,33), [185,184,171]
#> 171: (52,54,40), [153,172,170]
#> 172: (12,52,40), [171,158,165]
#> 173: (70,59,45), [183,169,166]
#> 174: (63,80,59), [175,166,164]
#> 175: (59,80,10), [167,0,174]
#> 176: (84,96,11), [182,178,177]
#> 177: (84,45,96), [179,176,180]
#> 178: (84,11,52), [185,168,176]
#> 179: (96,45,27), [183,181,177]
#> 180: (45,84,46), [168,159,177]
#> 181: (96,27,33), [0,182,179]
#> 182: (96,33,11), [185,176,181]
#> 183: (27,45,59), [173,0,179]
#> 184: (54,33,60), [0,152,170]
#> 185: (11,33,52), [170,178,182]
#> boundary nodes:  19 55 68 8 87 10 59 27 33 60 48 23 67 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.004667947 0.021564179 0.050933713 0.070893565 0.081632190 0.279969307 
#>                                                                         
#> 0.025200915 0.055864836 0.177829312 0.209195051 0.202378911 0.088284092 
#>                                                                         
#> 0.174678185 0.203550656 0.143222563 0.125850713 0.032953621 0.296613769 
#>                                                                         
#> 0.063612133 0.184357468 0.209214519 0.106140274 0.066669775 0.064265905 
#>                                                                         
#> 0.129811391 0.075104807 0.086023923 0.120356761 0.047126074 0.045620009 
#>                                                                         
#> 0.055416104 0.114797354 0.086970702 0.078319629 0.147748544 0.148144809 
#>                                                                         
#> 0.150155701 0.082506157 0.087863608 0.009979990 0.128877515 0.062681674 
#>                                                                         
#> 0.108189060 0.140533711 0.071194551 0.105631144 0.040844079 0.069525348 
#>                                                                         
#> 0.010164171 0.125947484 0.156707574 0.095724789 0.161398202 0.061015997 
#>                                                                         
#> 0.008331062 0.145770227 0.167923608 0.091222166 0.074105756 0.135483999 
#>                                                                         
#> 0.170323550 0.161076298 0.144374926 0.129546287 0.324137061 0.215341537 
#>                                                                         
#> 0.273185224 0.049484906 0.117111428 0.103335274 0.114964536 0.104085171 
#>                                                                         
#> 0.150544985 0.058807567 0.044052769 0.038858388 0.122209780 0.052378968 
#>                                                                         
#> 0.079416225 0.076134295 0.051692013 0.041654452 0.103074283 0.197072726 
#>                                                                         
#> 0.075376025 0.058901210 0.153286731 0.094943197 0.017900637 0.122439121 
#>                                                                         
#> 0.116082517 0.027551251 0.082504891 0.047941459 0.102813137 0.027492129 
#>                                                                         
#> 0.207049748 0.538722549 0.090017367 0.033614650 0.099523637 0.146581800 
#>                                                                         
#> 0.118285472 0.096839996 0.048515452 0.130647981 0.212305821 0.022089967 
#>                                                                         
#> 0.204595866 0.036525037 0.026093566 0.034630994 0.109355467 0.085504126 
#>                                                                         
#> 0.073480135 0.061864433 0.061102554 0.159062187 0.123143606 0.073350180 
#>                                                                         
#> 0.044653092 0.177699511 0.060702782 0.237207856 0.086842355 0.210110522 
#>                                                                         
#> 0.166853689 0.113546648 0.052560916 0.065828474 0.167080929 0.217474896 
#>                                                                         
#> 0.082855114 0.119999521 0.103711683 0.090571564 0.119887451 0.173399059 
#>                                                                         
#> 0.148566568 0.058808983 0.115135254 0.095053031 0.219340160 0.158074103 
#>                                                                         
#> 0.307461898 0.085899872 0.028905246 0.087728239 0.090323792 0.052026181 
#>                                                                         
#> 0.160435911 0.121410725 0.061016305 0.472876005 0.081187879 0.085686722 
#>                                                                         
#> 0.115424841 0.025377649 0.270889665 0.205045066 0.190422421 0.036990406 
#>                                                                         
#> 0.164212700 0.257050798 0.308534043 0.575703235 0.220115453 0.049575533 
#>                                                                         
#> 0.119601575 0.013559281 0.118013579 0.079731352 0.219887173 0.086767092 
#>                                                                         
#> 0.207693904 0.074503520 0.098668864 0.191017656 0.067478611 0.046383701 
#>                                                                         
#> 0.196550152 0.157728162 0.019694888 0.141957794 0.087748276 0.132942361 
#>                                                                         
#> 0.224367095 0.098042111 0.245067673 0.095698623 0.139761007 0.144883154 
#>                                                                         
#> 0.139232091 0.178960723 0.087242766 0.192029013 0.135986802 0.195538885 
#>                                                                         
#> 0.256426009 0.053899601 0.119507131 0.151739725 0.109614580 0.170811091 
#>                                                                         
#> 0.161163022 0.063536847 0.140417438 0.061677014 0.073858976 0.135074508 
#>                                                                         
#> 0.027378524 0.079622975 0.097841978 0.029097212 0.035613762 0.099577865 
#>                                                                         
#> 0.120824827 0.674701988 0.084558526 0.181332693 0.117115705 0.107945868 
#>                                                                         
#> 0.060639377 0.121358398 0.097594589 0.235638759 0.134117767 0.126204572 
#>                                                                         
#> 0.086146729 0.100536578 0.112936854 0.173769410 0.169236448 0.132985839 
#>                                                                         
#> 0.031839932 0.238674244 0.104186846 0.026682767 0.124897457 0.118939733 
#>                                                                         
#> 0.106570862 0.102682937 0.239607578 0.036670732 0.119033900 0.079705924 
#>                                                                         
#> 0.178241395 0.114785635 0.207702637 0.193533306 0.377024278 0.156203743 
#>                                                                         
#> 0.393286239 0.036696136 0.021559098 0.079722904 0.198994844 0.080987928 
#>                                                                         
#> 0.200035315 0.096390097 0.079339085 0.155345724 0.120176508 0.144936829 
#>                                                                         
#> 0.106702588 0.052237559 0.241164582 0.045348385 0.094925383 0.326296823 
#>                                                                         
#> 0.171692833 0.052951035 0.082761118 0.061375138 0.039204065 0.050701715 
#>                                                                         
#> 0.015791075 0.112810443 0.130348846 0.135786955 0.179866909 0.100800241 
#>                         
#> 0.283928573 0.092355531 
#> 
#> $alpha
#> [1] 0.2975994
#> 
#> attr(,"class")
#> [1] "scree"