Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1         my1
#>    28   29 0.97518698 0.05334485 0.81998224 0.08736483  0.898325583  0.07373529
#>    29   62 0.81998224 0.08736483 0.93995916 0.15589540  0.900642644  0.08543960
#>    66   86 0.88518196 0.19514401 0.79501915 0.15766587  0.851608756  0.14871915
#>    22   62 0.90625581 0.16592749 0.93995916 0.15589540  0.938033896  0.21105753
#>    39   63 0.98271470 0.17409100 0.99715196 0.21282317  0.941548092  0.21149248
#>    62   28 0.93995916 0.15589540 0.97518698 0.05334485  0.983272972  0.11344847
#>    86   29 0.79501915 0.15766587 0.81998224 0.08736483  0.852371565  0.13844848
#>    51   56 0.75962884 0.10443650 0.72606127 0.03368028  0.777555051  0.05259157
#>    63   66 0.99715196 0.21282317 0.88518196 0.19514401  0.939685985  0.21336326
#>    39   28 0.98271470 0.17409100 0.97518698 0.05334485  1.249028649  0.09688036
#>    56    1 0.72606127 0.03368028 0.96619843 0.03230197  0.845559482 -0.06638198
#>    72   51 0.58207110 0.04009406 0.75962884 0.10443650  0.628553241  0.18898624
#>    51   86 0.75962884 0.10443650 0.79501915 0.15766587  0.666809894  0.20452809
#>    97   89 0.92765684 0.33107350 0.77406513 0.29249244  0.859396671  0.27780229
#>    63   97 0.99715196 0.21282317 0.92765684 0.33107350  0.933150417  0.25475592
#>    72   95 0.58207110 0.04009406 0.63295689 0.00000000  0.656060524  0.08166036
#>    47   33 0.72007155 0.35972148 0.66352689 0.34099385  0.709756033  0.29614037
#>    87   47 0.74591167 0.41692044 0.72007155 0.35972148  0.791578084  0.36185402
#>    86   89 0.79501915 0.15766587 0.77406513 0.29249244  0.670821293  0.20740527
#>    81   68 0.92499056 0.43015262 0.86331926 0.47175342  0.845234603  0.37843090
#>    87   89 0.74591167 0.41692044 0.77406513 0.29249244  0.835771750  0.37185341
#>    16   81 0.97440031 0.50603072 0.92499056 0.43015262  1.077776510  0.38468875
#>    85   33 0.70849778 0.41509693 0.66352689 0.34099385  0.681777930  0.38061512
#>    87   85 0.74591167 0.41692044 0.70849778 0.41509693  0.728454287  0.39037074
#>    87   68 0.74591167 0.41692044 0.86331926 0.47175342  0.773593968  0.51075970
#>    82   59 0.47720665 0.09014170 0.28198314 0.01927814  0.442268028 -0.11794959
#>    82   72 0.47720665 0.09014170 0.58207110 0.04009406  0.572374920  0.15466240
#>    72  100 0.58207110 0.04009406 0.49239749 0.27486944  0.582633390  0.17482216
#>    44   19 0.67394681 0.48175034 0.61080703 0.43936138  0.661029272  0.43277255
#>    35   15 1.00000000 0.62499523 0.86762260 0.65292152  0.921308496  0.57969204
#>    63   16 0.99715196 0.21282317 0.97440031 0.50603072  1.130342429  0.37064467
#>    72   33 0.58207110 0.04009406 0.66352689 0.34099385  0.623381763  0.19038620
#>   100   33 0.49239749 0.27486944 0.66352689 0.34099385  0.587002234  0.28453610
#>    37   33 0.55461820 0.37394961 0.66352689 0.34099385  0.615446822  0.37853675
#>    68   44 0.86331926 0.47175342 0.67394681 0.48175034  0.770945912  0.52056490
#>    84   44 0.71570376 0.43188739 0.67394681 0.48175034  0.674367003  0.43968640
#>    48   73 0.48936538 0.35924609 0.39425225 0.27838647  0.444714221  0.31539873
#>   100   37 0.49239749 0.27486944 0.55461820 0.37394961  0.532815177  0.31856467
#>    48   37 0.48936538 0.35924609 0.55461820 0.37394961  0.520074587  0.37510623
#>    37   19 0.55461820 0.37394961 0.61080703 0.43936138  0.570740003  0.41693999
#>    48   13 0.48936538 0.35924609 0.52547560 0.40924460  0.511626127  0.38120793
#>    19   77 0.61080703 0.43936138 0.58801972 0.51943591  0.553324944  0.46628297
#>    77   67 0.58801972 0.51943591 0.50761733 0.41222070  0.528182262  0.48055385
#>    77   44 0.58801972 0.51943591 0.67394681 0.48175034  0.660658143  0.56825499
#>    14   90 0.38524986 0.26887061 0.47405096 0.09506107  0.443462030  0.18902234
#>    88   14 0.34429179 0.23415199 0.38524986 0.26887061  0.333609776  0.28827245
#>    73   14 0.39425225 0.27838647 0.38524986 0.26887061  0.441464240  0.22470576
#>    15   68 0.86762260 0.65292152 0.86331926 0.47175342  0.896422940  0.56160226
#>    88   83 0.34429179 0.23415199 0.24409072 0.10854570  0.360067325  0.11879692
#>    48    4 0.48936538 0.35924609 0.43460601 0.47410482  0.396326157  0.38537200
#>    44   34 0.67394681 0.48175034 0.77132072 0.67332273  0.743821128  0.56676726
#>    68   34 0.86331926 0.47175342 0.77132072 0.67332273  0.772612384  0.55213302
#>    18   74 0.62363815 0.78401721 0.67703234 0.65422899  0.700660118  0.73982649
#>    59   95 0.28198314 0.01927814 0.63295689 0.00000000  0.446840869 -0.18387291
#>    77   74 0.58801972 0.51943591 0.67703234 0.65422899  0.631797681  0.58731343
#>    70   92 0.29716087 0.24182175 0.26273978 0.18676656  0.313461256  0.19334276
#>    65   70 0.31161618 0.33886308 0.29716087 0.24182175  0.328657996  0.28672723
#>    48   67 0.48936538 0.35924609 0.50761733 0.41222070  0.439625476  0.40601513
#>    83   59 0.24409072 0.10854570 0.28198314 0.01927814  0.359243620  0.10474987
#>    70   46 0.29716087 0.24182175 0.29988220 0.33458056  0.270191174  0.28903230
#>    65   46 0.31161618 0.33886308 0.29988220 0.33458056  0.323727526  0.28746167
#>    77   25 0.58801972 0.51943591 0.62161515 0.66743947  0.584752586  0.59799223
#>    18   25 0.62363815 0.78401721 0.62161515 0.66743947  0.664619644  0.72499963
#>    92   83 0.26273978 0.18676656 0.24409072 0.10854570  0.353931854  0.12369143
#>     4   53 0.43460601 0.47410482 0.50680100 0.60825474  0.447570353  0.55362927
#>     4   77 0.43460601 0.47410482 0.58801972 0.51943591  0.503388124  0.52359002
#>    45   12 0.21108859 0.07144677 0.20575280 0.03298772  0.215191440  0.05127787
#>    49   83 0.23550380 0.05459842 0.24409072 0.10854570  0.236462976  0.08210279
#>    79   64 0.16410255 0.24681636 0.13377188 0.17617117  0.197502124  0.19064300
#>    58    9 0.07945663 0.38994588 0.22146698 0.26490444  0.193074782  0.37582099
#>    12   59 0.20575280 0.03298772 0.28198314 0.01927814  0.241801756  0.01464401
#>    65    4 0.31161618 0.33886308 0.43460601 0.47410482  0.383526871  0.39701177
#>    64   83 0.13377188 0.17617117 0.24409072 0.10854570  0.199987223  0.16039419
#>    92    9 0.26273978 0.18676656 0.22146698 0.26490444  0.207941867  0.20779123
#>     4   69 0.43460601 0.47410482 0.43297001 0.63286967  0.398812401  0.55312684
#>    69   53 0.43297001 0.63286967 0.50680100 0.60825474  0.503178907  0.72042374
#>    31   41 0.31433863 0.53338236 0.26884830 0.59929422  0.215043070  0.51350556
#>    76    4 0.29794962 0.43206817 0.43460601 0.47410482  0.359832203  0.47404045
#>    94   45 0.12766274 0.12993296 0.21108859 0.07144677  0.187118331  0.12599836
#>    79    9 0.16410255 0.24681636 0.22146698 0.26490444  0.165009426  0.34394699
#>    74   34 0.67703234 0.65422899 0.77132072 0.67332273  0.708292492  0.74221413
#>    27   20 0.63246345 0.79669843 0.56786845 0.84003653  0.596230924  0.81250237
#>    36   27 0.78824773 0.84723812 0.63246345 0.79669843  0.728757213  0.76524683
#>    64   94 0.13377188 0.17617117 0.12766274 0.12993296  0.190633770  0.14513571
#>    15   36 0.86762260 0.65292152 0.78824773 0.84723812  0.838836716  0.75453290
#>    10   32 0.39738117 0.68298254 0.34758848 0.64397331  0.390540572  0.64043097
#>    96   10 0.40989224 0.78772980 0.39738117 0.68298254  0.505100409  0.72323729
#>    32   17 0.34758848 0.64397331 0.35118291 0.66892593  0.303342511  0.66308214
#>    32   31 0.34758848 0.64397331 0.31433863 0.53338236  0.381821688  0.57338702
#>    76   58 0.29794962 0.43206817 0.07945663 0.38994588  0.194439223  0.38125323
#>    20   96 0.56786845 0.84003653 0.40989224 0.78772980  0.515837571  0.73246725
#>    27   57 0.63246345 0.79669843 0.75764962 0.87962162  0.707863470  0.81882585
#>    57   36 0.75764962 0.87962162 0.78824773 0.84723812  0.823338295  0.91104137
#>    43   79 0.08386502 0.28674531 0.16410255 0.24681636  0.161289555  0.34174708
#>    96   53 0.40989224 0.78772980 0.50680100 0.60825474  0.515777390  0.72900240
#>    43   38 0.08386502 0.28674531 0.03892695 0.30014130  0.046113453  0.24217666
#>    43   64 0.08386502 0.28674531 0.13377188 0.17617117  0.105707264  0.23005403
#>    36   50 0.78824773 0.84723812 0.92404741 0.86864439  0.850159525  0.89592895
#>    41   42 0.26884830 0.59929422 0.26134991 0.63942891  0.235661718  0.61386176
#>    41   32 0.26884830 0.59929422 0.34758848 0.64397331  0.305251095  0.62686318
#>    42   17 0.26134991 0.63942891 0.35118291 0.66892593  0.290965204  0.70077714
#>    17   10 0.35118291 0.66892593 0.39738117 0.68298254  0.360968607  0.71970998
#>    75   96 0.42253047 0.81058877 0.40989224 0.78772980  0.511131701  0.74667987
#>    26   99 0.32691957 0.75863866 0.20680089 0.68127334  0.272765240  0.71078778
#>    20   30 0.56786845 0.84003653 0.67315724 0.88932595  0.611946120  0.88298091
#>    76   31 0.29794962 0.43206817 0.31433863 0.53338236  0.204985083  0.49908918
#>     6   58 0.19941989 0.61355861 0.07945663 0.38994588  0.174873670  0.48274193
#>    75   11 0.42253047 0.81058877 0.46036333 0.84648345  0.413387865  0.85811023
#>    23   94 0.01214211 0.28968974 0.12766274 0.12993296  0.012930105  0.16861449
#>    43   58 0.08386502 0.28674531 0.07945663 0.38994588  0.074692576  0.33804793
#>    40   26 0.36986932 0.87927655 0.32691957 0.75863866  0.352922459  0.81734553
#>    11   20 0.46036333 0.84648345 0.56786845 0.84003653  0.515062122  0.85903880
#>    52    5 0.50508404 0.97617224 0.55293880 0.90045077  0.622535902  0.99741749
#>    30   57 0.67315724 0.88932595 0.75764962 0.87962162  0.730032884  1.01184761
#>    24   50 0.85973345 0.97410955 0.92404741 0.86864439  0.883440330  0.91622399
#>     3   80 0.89468131 0.97992150 0.89020410 0.97840925  0.910847963  0.92467447
#>     3   60 0.89468131 0.97992150 0.89095967 1.00000000  0.889033472  0.98925881
#>    35   60 1.00000000 0.62499523 0.89095967 1.00000000  4.325203118  1.79522144
#>    38   58 0.03892695 0.30014130 0.07945663 0.38994588 -0.005400561  0.37419474
#>     6   93 0.19941989 0.61355861 0.18157558 0.63841079  0.112799491  0.57019599
#>    99   93 0.20680089 0.68127334 0.18157558 0.63841079  0.218073964  0.64578492
#>    99    6 0.20680089 0.68127334 0.19941989 0.61355861  0.222529182  0.64529930
#>    71   99 0.15226143 0.77297966 0.20680089 0.68127334  0.239334612  0.76269272
#>    24   60 0.85973345 0.97410955 0.89095967 1.00000000  0.683821539  1.21805133
#>    75   40 0.42253047 0.81058877 0.36986932 0.87927655  0.369895928  0.82476609
#>    40   11 0.36986932 0.87927655 0.46036333 0.84648345  0.438232397  0.92666982
#>    11    5 0.46036333 0.84648345 0.55293880 0.90045077  0.485047002  0.91052669
#>    24   52 0.85973345 0.97410955 0.50508404 0.97617224  0.683221376  1.11486179
#>    23   58 0.01214211 0.28968974 0.07945663 0.38994588 -0.065331721  0.41443404
#>    93   61 0.18157558 0.63841079 0.12663381 0.69363916  0.094887082  0.60711466
#>    40    2 0.36986932 0.87927655 0.22245021 0.86027493  0.300512709  0.83600456
#>     8   93 0.02434053 0.53071913 0.18157558 0.63841079  0.094496354  0.59691946
#>    71   61 0.15226143 0.77297966 0.12663381 0.69363916  0.172160343  0.72274293
#>    55   71 0.18204532 0.83637525 0.15226143 0.77297966  0.239972878  0.77046613
#>    40   52 0.36986932 0.87927655 0.50508404 0.97617224  0.385842937  0.99977757
#>     2   55 0.22245021 0.86027493 0.18204532 0.83637525  0.243934532  0.77784928
#>     2   78 0.22245021 0.86027493 0.27666475 0.91617167  0.192660355  0.94340813
#>     7   54 0.01084818 0.64495387 0.00000000 0.66434801 -0.149256996  0.56812956
#>    54   23 0.00000000 0.66434801 0.01214211 0.28968974 -0.791326609  0.45117643
#>    71   54 0.15226143 0.77297966 0.00000000 0.66434801  0.043591577  0.76427169
#>    21   55 0.13947548 0.84509034 0.18204532 0.83637525  0.172582217  0.89847777
#>     2   91 0.22245021 0.86027493 0.26438461 0.99492699  0.185025426  0.94578586
#>    78   40 0.27666475 0.91617167 0.36986932 0.87927655  0.351132663  0.96811835
#>    52   60 0.50508404 0.97617224 0.89095967 1.00000000  0.559598512  3.22976505
#>    54   21 0.00000000 0.66434801 0.13947548 0.84509034 -0.011537411  0.81743770
#>    21    2 0.13947548 0.84509034 0.22245021 0.86027493  0.169284477  0.91649795
#>     1   29 0.96619843 0.03230197 0.81998224 0.08736483  0.898325583  0.07373529
#>    22   86 0.90625581 0.16592749 0.79501915 0.15766587  0.851608756  0.14871915
#>    66   62 0.88518196 0.19514401 0.93995916 0.15589540  0.938033896  0.21105753
#>    62   63 0.93995916 0.15589540 0.99715196 0.21282317  0.941548092  0.21149248
#>    22   29 0.90625581 0.16592749 0.81998224 0.08736483  0.852371565  0.13844848
#>    29   56 0.81998224 0.08736483 0.72606127 0.03368028  0.777555051  0.05259157
#>    63   28 0.99715196 0.21282317 0.97518698 0.05334485  1.249028649  0.09688036
#>    33   51 0.66352689 0.34099385 0.75962884 0.10443650  0.628553241  0.18898624
#>    33   86 0.66352689 0.34099385 0.79501915 0.15766587  0.666809894  0.20452809
#>    89   66 0.77406513 0.29249244 0.88518196 0.19514401  0.817724429  0.23023617
#>    66   97 0.88518196 0.19514401 0.92765684 0.33107350  0.933150417  0.25475592
#>    95   56 0.63295689 0.00000000 0.72606127 0.03368028  0.843436195 -0.43631326
#>    56   72 0.72606127 0.03368028 0.58207110 0.04009406  0.657306771  0.10963874
#>    89   33 0.77406513 0.29249244 0.66352689 0.34099385  0.709756033  0.29614037
#>    89   47 0.77406513 0.29249244 0.72007155 0.35972148  0.791578084  0.36185402
#>    97   68 0.92765684 0.33107350 0.86331926 0.47175342  0.845234603  0.37843090
#>    97   81 0.92765684 0.33107350 0.92499056 0.43015262  1.077776510  0.38468875
#>    47   85 0.72007155 0.35972148 0.70849778 0.41509693  0.728454287  0.39037074
#>    84   68 0.71570376 0.43188739 0.86331926 0.47175342  0.773593968  0.51075970
#>    97   87 0.92765684 0.33107350 0.74591167 0.41692044  0.837127263  0.37472315
#>    19   85 0.61080703 0.43936138 0.70849778 0.41509693  0.652489213  0.39838954
#>    84   87 0.71570376 0.43188739 0.74591167 0.41692044  0.727154879  0.41703137
#>    72   59 0.58207110 0.04009406 0.28198314 0.01927814  0.442268028 -0.11794959
#>    90   72 0.47405096 0.09506107 0.58207110 0.04009406  0.572374920  0.15466240
#>    90  100 0.47405096 0.09506107 0.49239749 0.27486944  0.582633390  0.17482216
#>    16   15 0.97440031 0.50603072 0.86762260 0.65292152  0.921308496  0.57969204
#>    97   16 0.92765684 0.33107350 0.97440031 0.50603072  1.130342429  0.37064467
#>    19   33 0.61080703 0.43936138 0.66352689 0.34099385  0.615446822  0.37853675
#>    85   44 0.70849778 0.41509693 0.67394681 0.48175034  0.674367003  0.43968640
#>   100   73 0.49239749 0.27486944 0.39425225 0.27838647  0.444714221  0.31539873
#>    13   37 0.52547560 0.40924460 0.55461820 0.37394961  0.520074587  0.37510623
#>    13   19 0.52547560 0.40924460 0.61080703 0.43936138  0.570740003  0.41693999
#>    67   13 0.50761733 0.41222070 0.52547560 0.40924460  0.511626127  0.38120793
#>    13   77 0.52547560 0.40924460 0.58801972 0.51943591  0.553324944  0.46628297
#>    74   44 0.67703234 0.65422899 0.67394681 0.48175034  0.660658143  0.56825499
#>    65   48 0.31161618 0.33886308 0.48936538 0.35924609  0.398056712  0.37028077
#>    65   14 0.31161618 0.33886308 0.38524986 0.26887061  0.333609776  0.28827245
#>   100   14 0.49239749 0.27486944 0.38524986 0.26887061  0.441464240  0.22470576
#>    16   68 0.97440031 0.50603072 0.86331926 0.47175342  0.896422940  0.56160226
#>    35   63 1.00000000 0.62499523 0.99715196 0.21282317  1.690708611  0.41412668
#>    90   83 0.47405096 0.09506107 0.24409072 0.10854570  0.360067325  0.11879692
#>    65   73 0.31161618 0.33886308 0.39425225 0.27838647  0.352862429  0.30852669
#>    34   15 0.77132072 0.67332273 0.86762260 0.65292152  0.798457998  0.56392925
#>    27   74 0.63246345 0.79669843 0.67703234 0.65422899  0.700660118  0.73982649
#>    90   82 0.47405096 0.09506107 0.47720665 0.09014170  0.386576597  0.03547589
#>    25   74 0.62161515 0.66743947 0.67703234 0.65422899  0.631797681  0.58731343
#>    88   92 0.34429179 0.23415199 0.26273978 0.18676656  0.313461256  0.19334276
#>    88   70 0.34429179 0.23415199 0.29716087 0.24182175  0.328657996  0.28672723
#>     4   67 0.43460601 0.47410482 0.50761733 0.41222070  0.439625476  0.40601513
#>    90   59 0.47405096 0.09506107 0.28198314 0.01927814  0.359243620  0.10474987
#>     9   46 0.22146698 0.26490444 0.29988220 0.33458056  0.270191174  0.28903230
#>    76   65 0.29794962 0.43206817 0.31161618 0.33886308  0.288813886  0.38312410
#>    53   25 0.50680100 0.60825474 0.62161515 0.66743947  0.584752586  0.59799223
#>    53   77 0.50680100 0.60825474 0.58801972 0.51943591  0.503388124  0.52359002
#>    49   12 0.23550380 0.05459842 0.20575280 0.03298772  0.215191440  0.05127787
#>    45   83 0.21108859 0.07144677 0.24409072 0.10854570  0.236462976  0.08210279
#>    92   64 0.26273978 0.18676656 0.13377188 0.17617117  0.197502124  0.19064300
#>    49   59 0.23550380 0.05459842 0.28198314 0.01927814  0.241801756  0.01464401
#>     9   70 0.22146698 0.26490444 0.29716087 0.24182175  0.252612145  0.23138628
#>    31   69 0.31433863 0.53338236 0.43297001 0.63286967  0.398812401  0.55312684
#>    10   53 0.39738117 0.68298254 0.50680100 0.60825474  0.503178907  0.72042374
#>     6   41 0.19941989 0.61355861 0.26884830 0.59929422  0.215043070  0.51350556
#>    31    4 0.31433863 0.53338236 0.43460601 0.47410482  0.359832203  0.47404045
#>    53   18 0.50680100 0.60825474 0.62363815 0.78401721  0.517970623  0.72754446
#>    27   34 0.63246345 0.79669843 0.77132072 0.67332273  0.708292492  0.74221413
#>    50   35 0.92404741 0.86864439 1.00000000 0.62499523  0.956181518  0.74499862
#>    18   20 0.62363815 0.78401721 0.56786845 0.84003653  0.596230924  0.81250237
#>    83   94 0.24409072 0.10854570 0.12766274 0.12993296  0.190633770  0.14513571
#>    34   36 0.77132072 0.67332273 0.78824773 0.84723812  0.838836716  0.75453290
#>    69   32 0.43297001 0.63286967 0.34758848 0.64397331  0.390540572  0.64043097
#>    50   15 0.92404741 0.86864439 0.86762260 0.65292152  0.870414667  0.76743194
#>    46   58 0.29988220 0.33458056 0.07945663 0.38994588  0.194439223  0.38125323
#>    18   96 0.62363815 0.78401721 0.40989224 0.78772980  0.515837571  0.73246725
#>    24   36 0.85973345 0.97410955 0.78824773 0.84723812  0.823338295  0.91104137
#>    58   79 0.07945663 0.38994588 0.16410255 0.24681636  0.161289555  0.34174708
#>    23   38 0.01214211 0.28968974 0.03892695 0.30014130  0.046113453  0.24217666
#>    30   27 0.67315724 0.88932595 0.63246345 0.79669843  0.624921053  0.85526471
#>     6   42 0.19941989 0.61355861 0.26134991 0.63942891  0.235661718  0.61386176
#>    42   32 0.26134991 0.63942891 0.34758848 0.64397331  0.305251095  0.62686318
#>    26   17 0.32691957 0.75863866 0.35118291 0.66892593  0.290965204  0.70077714
#>    26   10 0.32691957 0.75863866 0.39738117 0.68298254  0.360968607  0.71970998
#>    42   99 0.26134991 0.63942891 0.20680089 0.68127334  0.272765240  0.71078778
#>     5   30 0.55293880 0.90045077 0.67315724 0.88932595  0.611946120  0.88298091
#>     6   31 0.19941989 0.61355861 0.31433863 0.53338236  0.204985083  0.49908918
#>    26   96 0.32691957 0.75863866 0.40989224 0.78772980  0.380708679  0.73809471
#>    20   75 0.56786845 0.84003653 0.42253047 0.81058877  0.508965315  0.75737196
#>     5   20 0.55293880 0.90045077 0.56786845 0.84003653  0.515062122  0.85903880
#>    24   57 0.85973345 0.97410955 0.75764962 0.87962162  0.730032884  1.01184761
#>    80   50 0.89020410 0.97840925 0.92404741 0.86864439  0.883440330  0.91622399
#>    80   60 0.89020410 0.97840925 0.89095967 1.00000000  0.889033472  0.98925881
#>    50   60 0.92404741 0.86864439 0.89095967 1.00000000  4.325203118  1.79522144
#>    23   43 0.01214211 0.28968974 0.08386502 0.28674531  0.044481333  0.20242013
#>     8    6 0.02434053 0.53071913 0.19941989 0.61355861  0.147138523  0.49762122
#>    80   24 0.89020410 0.97840925 0.85973345 0.97410955  0.873055534  0.98981796
#>    50    3 0.92404741 0.86864439 0.89468131 0.97992150  1.784572830  1.15525100
#>    96   40 0.40989224 0.78772980 0.36986932 0.87927655  0.369895928  0.82476609
#>    52   11 0.50508404 0.97617224 0.46036333 0.84648345  0.438232397  0.92666982
#>    30   52 0.67315724 0.88932595 0.50508404 0.97617224  0.683221376  1.11486179
#>     8   58 0.02434053 0.53071913 0.07945663 0.38994588 -0.065331721  0.41443404
#>     7   61 0.01084818 0.64495387 0.12663381 0.69363916  0.094887082  0.60711466
#>    61   99 0.12663381 0.69363916 0.20680089 0.68127334  0.165098574  0.67696180
#>    26    2 0.32691957 0.75863866 0.22245021 0.86027493  0.300512709  0.83600456
#>     7   93 0.01084818 0.64495387 0.18157558 0.63841079  0.094496354  0.59691946
#>    26   71 0.32691957 0.75863866 0.15226143 0.77297966  0.239972878  0.77046613
#>    91   52 0.26438461 0.99492699 0.50508404 0.97617224  0.385842937  0.99977757
#>    26   55 0.32691957 0.75863866 0.18204532 0.83637525  0.243934532  0.77784928
#>    91   78 0.26438461 0.99492699 0.27666475 0.91617167  0.192660355  0.94340813
#>     8   54 0.02434053 0.53071913 0.00000000 0.66434801 -0.149256996  0.56812956
#>     8   23 0.02434053 0.53071913 0.01214211 0.28968974 -0.791326609  0.45117643
#>    61   54 0.12663381 0.69363916 0.00000000 0.66434801  0.043591577  0.76427169
#>    21   71 0.13947548 0.84509034 0.15226143 0.77297966  0.026726731  0.78790997
#>    98   91 0.09305958 0.92981666 0.26438461 0.99492699  0.185025426  0.94578586
#>    91   40 0.26438461 0.99492699 0.36986932 0.87927655  0.351132663  0.96811835
#>    91   60 0.26438461 0.99492699 0.89095967 1.00000000  0.559598512  3.22976505
#>    98   21 0.09305958 0.92981666 0.13947548 0.84509034 -0.011537411  0.81743770
#>    98    2 0.09305958 0.92981666 0.22245021 0.86027493  0.169284477  0.91649795
#>     1   28 0.96619843 0.03230197 0.97518698 0.05334485  0.898325583  0.07373529
#>    22   66 0.90625581 0.16592749 0.88518196 0.19514401  0.851608756  0.14871915
#>    62   39 0.93995916 0.15589540 0.98271470 0.17409100  0.941548092  0.21149248
#>    29   51 0.81998224 0.08736483 0.75962884 0.10443650  0.777555051  0.05259157
#>    95    1 0.63295689 0.00000000 0.96619843 0.03230197  0.843436195 -0.43631326
#>    84   85 0.71570376 0.43188739 0.70849778 0.41509693  0.727154879  0.41703137
#>    16   35 0.97440031 0.50603072 1.00000000 0.62499523  0.921308496  0.57969204
#>   100   48 0.49239749 0.27486944 0.48936538 0.35924609  0.444714221  0.31539873
#>    65   88 0.31161618 0.33886308 0.34429179 0.23415199  0.333609776  0.28827245
#>    90   88 0.47405096 0.09506107 0.34429179 0.23415199  0.360067325  0.11879692
#>    27   18 0.63246345 0.79669843 0.62363815 0.78401721  0.700660118  0.73982649
#>    76   46 0.29794962 0.43206817 0.29988220 0.33458056  0.288813886  0.38312410
#>    49   45 0.23550380 0.05459842 0.21108859 0.07144677  0.215191440  0.05127787
#>    92   79 0.26273978 0.18676656 0.16410255 0.24681636  0.197502124  0.19064300
#>    10   69 0.39738117 0.68298254 0.43297001 0.63286967  0.503178907  0.72042374
#>    94   12 0.12766274 0.12993296 0.20575280 0.03298772  0.141866939  0.06145088
#>    26   42 0.32691957 0.75863866 0.26134991 0.63942891  0.290965204  0.70077714
#>     6   76 0.19941989 0.61355861 0.29794962 0.43206817  0.204985083  0.49908918
#>    64   23 0.13377188 0.17617117 0.01214211 0.28968974  0.012930105  0.16861449
#>    24   30 0.85973345 0.97410955 0.67315724 0.88932595  0.730032884  1.01184761
#>     7    8 0.01084818 0.64495387 0.02434053 0.53071913  0.094496354  0.59691946
#>    98   54 0.09305958 0.92981666 0.00000000 0.66434801 -0.011537411  0.81743770
#>           mx2          my2 bp1 bp2
#>   0.900822198   0.08512526   0   0
#>   0.900822198   0.08512526   0   0
#>   0.817724429   0.23023617   0   0
#>   0.900642644   0.08543960   0   0
#>   1.249028649   0.09688036   0   0
#>   0.900822198   0.08512526   0   0
#>   0.796198625   0.11850212   0   0
#>   0.657306771   0.10963874   0   0
#>   0.933150417   0.25475592   0   0
#>   0.983272972   0.11344847   0   0
#>   0.843436195  -0.43631326   0   0
#>   0.657306771   0.10963874   0   0
#>   0.796198625   0.11850212   0   0
#>   0.835771750   0.37185341   0   0
#>   1.130342429   0.37064467   0   0
#>   0.446840869  -0.18387291   0   0
#>   0.681777930   0.38061512   0   0
#>   0.728454287   0.39037074   0   0
#>   0.817724429   0.23023617   0   0
#>   0.919138245   0.48798968   0   0
#>   0.791578084   0.36185402   0   0
#>   0.919138245   0.48798968   0   0
#>   0.652489213   0.39838954   0   0
#>   0.727154879   0.41703137   0   0
#>   0.837127263   0.37472315   0   0
#>   0.386576597   0.03547589   0   0
#>   0.442268028  -0.11794959   0   0
#>   0.623381763   0.19038620   0   0
#>   0.624864368   0.48664139   0   0
#>   0.956181518   0.74499862   0   0
#>   1.690708611   0.41412668   0   0
#>   0.628553241   0.18898624   0   0
#>   0.623381763   0.19038620   0   0
#>   0.587002234   0.28453610   0   0
#>   0.772612384   0.55213302   0   0
#>   0.770945912   0.52056490   0   0
#>   0.398056712   0.37028077   0   0
#>   0.587002234   0.28453610   0   0
#>   0.532815177   0.31856467   0   0
#>   0.615446822   0.37853675   0   0
#>   0.520074587   0.37510623   0   0
#>   0.624864368   0.48664139   0   0
#>   0.512666463   0.49218940   0   0
#>   0.624864368   0.48664139   0   0
#>   0.425513507   0.17985226   0   0
#>   0.425513507   0.17985226   0   0
#>   0.352862429   0.30852669   0   0
#>   0.798457998   0.56392925   0   0
#>   0.353931854   0.12369143   0   0
#>   0.439625476   0.40601513   0   0
#>   0.772612384   0.55213302   0   0
#>   0.798457998   0.56392925   0   0
#>   0.664619644   0.72499963   0   0
#>   0.290811083  -3.02451854   0   1
#>   0.660658143   0.56825499   0   0
#>   0.252612145   0.23138628   0   0
#>   0.323727526   0.28746167   0   0
#>   0.511626127   0.38120793   0   0
#>   0.286956964   0.07406553   0   0
#>   0.323727526   0.28746167   0   0
#>   0.288813886   0.38312410   0   0
#>   0.631797681   0.58731343   0   0
#>   0.517970623   0.72754446   0   0
#>   0.199987223   0.16039419   0   0
#>   0.503388124   0.52359002   0   0
#>   0.512666463   0.49218940   0   0
#>   0.141866939   0.06145088   0   0
#>   0.286956964   0.07406553   0   0
#>   0.105707264   0.23005403   0   0
#>   0.165009426   0.34394699   0   0
#>  -2.043502261 -12.69249333   0   1
#>   0.396326157   0.38537200   0   0
#>   0.190633770   0.14513571   0   0
#>   0.252612145   0.23138628   0   0
#>   0.447570353   0.55362927   0   0
#>   0.447570353   0.55362927   0   0
#>   0.326086329   0.59014420   0   0
#>   0.383526871   0.39701177   0   0
#>   0.141866939   0.06145088   0   0
#>   0.207941867   0.20779123   0   0
#>   0.743821128   0.56676726   0   0
#>   0.624921053   0.85526471   0   0
#>   0.714745789   0.80843586   0   0
#>   0.012930105   0.16861449   0   0
#>   0.870414667   0.76743194   0   0
#>   0.381629903   0.65180484   0   0
#>   0.380708679   0.73809471   0   0
#>   0.381629903   0.65180484   0   0
#>   0.326086329   0.59014420   0   0
#>   0.174873670   0.48274193   0   0
#>   0.511131701   0.74667987   0   0
#>   0.714745789   0.80843586   0   0
#>   0.714745789   0.80843586   0   0
#>   0.105707264   0.23005403   0   0
#>   0.505100409   0.72323729   0   0
#>   0.074692576   0.33804793   0   0
#>   0.044481333   0.20242013   0   0
#>   0.870414667   0.76743194   0   0
#>   0.305251095   0.62686318   0   0
#>   0.326086329   0.59014420   0   0
#>   0.303342511   0.66308214   0   0
#>   0.381629903   0.65180484   0   0
#>   0.369895928   0.82476609   0   0
#>   0.239334612   0.76269272   0   0
#>   0.624921053   0.85526471   0   0
#>   0.359832203   0.47404045   0   0
#>   0.147138523   0.49762122   0   0
#>   0.508965315   0.75737196   0   0
#>  -4.097389042  -2.80357022   0   1
#>   0.161289555   0.34174708   0   0
#>   0.300512709   0.83600456   0   0
#>   0.508965315   0.75737196   0   0
#>   0.485047002   0.91052669   0   0
#>   0.707863470   0.81882585   0   0
#>   0.850159525   0.89592895   0   0
#>   0.889033472   0.98925881   0   0
#>   1.784572830   1.15525100   0   0
#>   6.783954589   2.51015383   0   1
#>   0.074692576   0.33804793   0   0
#>   0.218073964   0.64578492   0   0
#>   0.165098574   0.67696180   0   0
#>   0.218073964   0.64578492   0   0
#>   0.172160343   0.72274293   0   0
#>   0.873055534   0.98981796   0   0
#>   0.413387865   0.85811023   0   0
#>   0.413387865   0.85811023   0   0
#>   0.515062122   0.85903880   0   0
#>   0.683821539   1.21805133   0   0
#>  -0.005400561   0.37419474   0   0
#>   0.165098574   0.67696180   0   0
#>   0.300100680   0.83920118   0   0
#>   0.112799491   0.57019599   0   0
#>   0.043591577   0.76427169   0   0
#>   0.154587580   0.81058099   0   0
#>   0.438232397   0.92666982   0   0
#>   0.172582217   0.89847777   0   0
#>   0.300100680   0.83920118   0   0
#>   0.061670273   0.68611243   0   0
#>  -3.457625107   0.36476573   0   1
#>   0.026726731   0.78790997   0   0
#>   0.154587580   0.81058099   0   0
#>   0.192660355   0.94340813   0   0
#>   0.300100680   0.83920118   0   0
#>   0.683821539   1.21805133   0   0
#>   0.026726731   0.78790997   0   0
#>   0.172582217   0.89847777   0   0
#>   0.845559482  -0.06638198   0   0
#>   0.852371565   0.13844848   0   0
#>   0.939685985   0.21336326   0   0
#>   0.939685985   0.21336326   0   0
#>   0.900642644   0.08543960   0   0
#>   0.845559482  -0.06638198   0   0
#>   7.402740095  -0.75067167   0   1
#>   0.666809894   0.20452809   0   0
#>   0.670821293   0.20740527   0   0
#>   0.859396671   0.27780229   0   0
#>   0.859396671   0.27780229   0   0
#>   0.656060524   0.08166036   0   0
#>   0.656060524   0.08166036   0   0
#>   0.670821293   0.20740527   0   0
#>   0.709756033   0.29614037   0   0
#>   0.837127263   0.37472315   0   0
#>   0.845234603   0.37843090   0   0
#>   0.681777930   0.38061512   0   0
#>   0.770945912   0.52056490   0   0
#>   0.835771750   0.37185341   0   0
#>   0.661029272   0.43277255   0   0
#>   0.773593968   0.51075970   0   0
#>   0.446840869  -0.18387291   0   0
#>   0.582633390   0.17482216   0   0
#>   0.443462030   0.18902234   0   0
#>   0.896422940   0.56160226   0   0
#>   1.077776510   0.38468875   0   0
#>   0.652489213   0.39838954   0   0
#>   0.661029272   0.43277255   0   0
#>   0.441464240   0.22470576   0   0
#>   0.570740003   0.41693999   0   0
#>   0.553324944   0.46628297   0   0
#>   0.528182262   0.48055385   0   0
#>   0.528182262   0.48055385   0   0
#>   0.743821128   0.56676726   0   0
#>   0.396326157   0.38537200   0   0
#>   0.352862429   0.30852669   0   0
#>   0.443462030   0.18902234   0   0
#>   0.919138245   0.48798968   0   0
#>   4.116764043   0.39736304   0   1
#>   0.359243620   0.10474987   0   0
#>   0.398056712   0.37028077   0   0
#>   0.838836716   0.75453290   0   0
#>   0.708292492   0.74221413   0   0
#>   0.572374920   0.15466240   0   0
#>   0.664619644   0.72499963   0   0
#>   0.353931854   0.12369143   0   0
#>   0.313461256   0.19334276   0   0
#>   0.512666463   0.49218940   0   0
#>   0.386576597   0.03547589   0   0
#>   0.193074782   0.37582099   0   0
#>   0.383526871   0.39701177   0   0
#>   0.517970623   0.72754446   0   0
#>   0.584752586   0.59799223   0   0
#>   0.241801756   0.01464401   0   0
#>   0.187118331   0.12599836   0   0
#>   0.199987223   0.16039419   0   0
#>   0.286956964   0.07406553   0   0
#>   0.270191174   0.28903230   0   0
#>   0.381821688   0.57338702   0   0
#>   0.505100409   0.72323729   0   0
#>   0.235661718   0.61386176   0   0
#>   0.398812401   0.55312684   0   0
#>   0.515777390   0.72900240   0   0
#>   0.728757213   0.76524683   0   0
#>   4.325203118   1.79522144   0   0
#>   0.515837571   0.73246725   0   0
#>   0.187118331   0.12599836   0   0
#>   0.728757213   0.76524683   0   0
#>   0.381821688   0.57338702   0   0
#>   0.956181518   0.74499862   0   0
#>   0.193074782   0.37582099   0   0
#>   0.515777390   0.72900240   0   0
#>   0.850159525   0.89592895   0   0
#>   0.165009426   0.34394699   0   0
#>  -0.005400561   0.37419474   0   0
#>   0.707863470   0.81882585   0   0
#>   0.222529182   0.64529930   0   0
#>   0.303342511   0.66308214   0   0
#>   0.360968607   0.71970998   0   0
#>   0.380708679   0.73809471   0   0
#>   0.222529182   0.64529930   0   0
#>   0.622535902   0.99741749   0   0
#>   0.215043070   0.51350556   0   0
#>   0.352922459   0.81734553   0   0
#>   0.511131701   0.74667987   0   0
#>   0.611946120   0.88298091   0   0
#>   0.823338295   0.91104137   0   0
#>   0.910847963   0.92467447   0   0
#>   0.873055534   0.98981796   0   0
#>   1.784572830   1.15525100   0   0
#>   0.046113453   0.24217666   0   0
#>   0.112799491   0.57019599   0   0
#>   0.883440330   0.91622399   0   0
#>   0.910847963   0.92467447   0   0
#>   0.352922459   0.81734553   0   0
#>   0.485047002   0.91052669   0   0
#>   0.622535902   0.99741749   0   0
#>   0.147138523   0.49762122   0   0
#>   0.061670273   0.68611243   0   0
#>   0.172160343   0.72274293   0   0
#>   0.243934532   0.77784928   0   0
#>   0.094887082   0.60711466   0   0
#>   0.239334612   0.76269272   0   0
#>   0.559598512   3.22976505   0   0
#>   0.239972878   0.77046613   0   0
#>   0.351132663   0.96811835   0   0
#>  -0.791326609   0.45117643   0   0
#>  -0.065331721   0.41443404   0   0
#>   0.061670273   0.68611243   0   0
#>   0.154587580   0.81058099   0   0
#>  -1.753265232   6.04601689   0   1
#>   0.385842937   0.99977757   0   0
#>   0.546677665   4.82563842   0   1
#>   0.169284477   0.91649795   0   0
#>   0.185025426   0.94578586   0   0
#>  41.087416888 -17.09318108   0   1
#>   0.938033896   0.21105753   0   0
#>   0.983272972   0.11344847   0   0
#>   0.796198625   0.11850212   0   0
#>   1.131606469  -3.40920648   0   1
#>   0.674367003   0.43968640   0   0
#>   1.690708611   0.41412668   0   0
#>   0.532815177   0.31856467   0   0
#>   0.328657996   0.28672723   0   0
#>   0.425513507   0.17985226   0   0
#>   0.596230924   0.81250237   0   0
#>   0.194439223   0.38125323   0   0
#>   0.236462976   0.08210279   0   0
#>   0.207941867   0.20779123   0   0
#>   0.390540572   0.64043097   0   0
#>  -6.114102633  -4.97777575   0   1
#>   0.272765240   0.71078778   0   0
#>   0.174873670   0.48274193   0   0
#>   0.044481333   0.20242013   0   0
#>   0.683221376   1.11486179   0   0
#>  -0.149256996   0.56812956   0   0
#>  -3.366223030   1.99341698   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.96619843 0.03230197
#>   [2,] 0.22245021 0.86027493
#>   [3,] 0.89468131 0.97992150
#>   [4,] 0.43460601 0.47410482
#>   [5,] 0.55293880 0.90045077
#>   [6,] 0.19941989 0.61355861
#>   [7,] 0.01084818 0.64495387
#>   [8,] 0.02434053 0.53071913
#>   [9,] 0.22146698 0.26490444
#>  [10,] 0.39738117 0.68298254
#>  [11,] 0.46036333 0.84648345
#>  [12,] 0.20575280 0.03298772
#>  [13,] 0.52547560 0.40924460
#>  [14,] 0.38524986 0.26887061
#>  [15,] 0.86762260 0.65292152
#>  [16,] 0.97440031 0.50603072
#>  [17,] 0.35118291 0.66892593
#>  [18,] 0.62363815 0.78401721
#>  [19,] 0.61080703 0.43936138
#>  [20,] 0.56786845 0.84003653
#>  [21,] 0.13947548 0.84509034
#>  [22,] 0.90625581 0.16592749
#>  [23,] 0.01214211 0.28968974
#>  [24,] 0.85973345 0.97410955
#>  [25,] 0.62161515 0.66743947
#>  [26,] 0.32691957 0.75863866
#>  [27,] 0.63246345 0.79669843
#>  [28,] 0.97518698 0.05334485
#>  [29,] 0.81998224 0.08736483
#>  [30,] 0.67315724 0.88932595
#>  [31,] 0.31433863 0.53338236
#>  [32,] 0.34758848 0.64397331
#>  [33,] 0.66352689 0.34099385
#>  [34,] 0.77132072 0.67332273
#>  [35,] 1.00000000 0.62499523
#>  [36,] 0.78824773 0.84723812
#>  [37,] 0.55461820 0.37394961
#>  [38,] 0.03892695 0.30014130
#>  [39,] 0.98271470 0.17409100
#>  [40,] 0.36986932 0.87927655
#>  [41,] 0.26884830 0.59929422
#>  [42,] 0.26134991 0.63942891
#>  [43,] 0.08386502 0.28674531
#>  [44,] 0.67394681 0.48175034
#>  [45,] 0.21108859 0.07144677
#>  [46,] 0.29988220 0.33458056
#>  [47,] 0.72007155 0.35972148
#>  [48,] 0.48936538 0.35924609
#>  [49,] 0.23550380 0.05459842
#>  [50,] 0.92404741 0.86864439
#>  [51,] 0.75962884 0.10443650
#>  [52,] 0.50508404 0.97617224
#>  [53,] 0.50680100 0.60825474
#>  [54,] 0.00000000 0.66434801
#>  [55,] 0.18204532 0.83637525
#>  [56,] 0.72606127 0.03368028
#>  [57,] 0.75764962 0.87962162
#>  [58,] 0.07945663 0.38994588
#>  [59,] 0.28198314 0.01927814
#>  [60,] 0.89095967 1.00000000
#>  [61,] 0.12663381 0.69363916
#>  [62,] 0.93995916 0.15589540
#>  [63,] 0.99715196 0.21282317
#>  [64,] 0.13377188 0.17617117
#>  [65,] 0.31161618 0.33886308
#>  [66,] 0.88518196 0.19514401
#>  [67,] 0.50761733 0.41222070
#>  [68,] 0.86331926 0.47175342
#>  [69,] 0.43297001 0.63286967
#>  [70,] 0.29716087 0.24182175
#>  [71,] 0.15226143 0.77297966
#>  [72,] 0.58207110 0.04009406
#>  [73,] 0.39425225 0.27838647
#>  [74,] 0.67703234 0.65422899
#>  [75,] 0.42253047 0.81058877
#>  [76,] 0.29794962 0.43206817
#>  [77,] 0.58801972 0.51943591
#>  [78,] 0.27666475 0.91617167
#>  [79,] 0.16410255 0.24681636
#>  [80,] 0.89020410 0.97840925
#>  [81,] 0.92499056 0.43015262
#>  [82,] 0.47720665 0.09014170
#>  [83,] 0.24409072 0.10854570
#>  [84,] 0.71570376 0.43188739
#>  [85,] 0.70849778 0.41509693
#>  [86,] 0.79501915 0.15766587
#>  [87,] 0.74591167 0.41692044
#>  [88,] 0.34429179 0.23415199
#>  [89,] 0.77406513 0.29249244
#>  [90,] 0.47405096 0.09506107
#>  [91,] 0.26438461 0.99492699
#>  [92,] 0.26273978 0.18676656
#>  [93,] 0.18157558 0.63841079
#>  [94,] 0.12766274 0.12993296
#>  [95,] 0.63295689 0.00000000
#>  [96,] 0.40989224 0.78772980
#>  [97,] 0.92765684 0.33107350
#>  [98,] 0.09305958 0.92981666
#>  [99,] 0.20680089 0.68127334
#> [100,] 0.49239749 0.27486944
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,28,29), [2,13,0]
#> 2: (62,29,28), [1,7,3]
#> 3: (22,29,62), [2,5,8]
#> 4: (22,66,86), [16,8,5]
#> 5: (66,22,62), [3,11,4]
#> 6: (62,39,63), [12,11,7]
#> 7: (39,62,28), [2,12,6]
#> 8: (22,86,29), [10,3,4]
#> 9: (29,51,56), [20,13,10]
#> 10: (51,29,86), [8,15,9]
#> 11: (62,63,66), [18,5,6]
#> 12: (63,39,28), [7,0,6]
#> 13: (29,56,1), [19,1,9]
#> 14: (33,72,51), [20,15,41]
#> 15: (33,51,86), [10,24,14]
#> 16: (89,86,66), [4,17,24]
#> 17: (66,97,89), [26,16,18]
#> 18: (66,63,97), [40,17,11]
#> 19: (95,1,56), [13,21,0]
#> 20: (56,51,72), [14,21,9]
#> 21: (56,72,95), [70,19,20]
#> 22: (89,47,33), [28,24,23]
#> 23: (89,87,47), [29,22,26]
#> 24: (33,86,89), [16,22,15]
#> 25: (97,81,68), [39,31,27]
#> 26: (97,87,89), [23,17,31]
#> 27: (97,16,81), [39,25,40]
#> 28: (47,85,33), [32,22,29]
#> 29: (47,87,85), [33,28,23]
#> 30: (84,87,68), [31,44,33]
#> 31: (97,68,87), [30,26,25]
#> 32: (19,33,85), [28,37,43]
#> 33: (84,85,87), [29,30,45]
#> 34: (72,82,59), [71,70,35]
#> 35: (90,82,72), [34,36,71]
#> 36: (90,72,100), [41,55,35]
#> 37: (85,44,19), [50,32,45]
#> 38: (16,35,15), [103,60,61]
#> 39: (16,68,81), [25,27,60]
#> 40: (97,63,16), [61,27,18]
#> 41: (100,72,33), [14,42,36]
#> 42: (37,100,33), [41,43,47]
#> 43: (19,37,33), [42,32,49]
#> 44: (84,68,44), [68,45,30]
#> 45: (85,84,44), [44,37,33]
#> 46: (100,48,73), [56,58,47]
#> 47: (48,100,37), [42,48,46]
#> 48: (13,48,37), [47,49,51]
#> 49: (13,37,19), [43,52,48]
#> 50: (77,19,44), [37,54,52]
#> 51: (67,48,13), [48,53,75]
#> 52: (13,19,77), [50,53,49]
#> 53: (13,77,67), [59,51,52]
#> 54: (74,77,44), [50,67,72]
#> 55: (100,14,90), [63,36,58]
#> 56: (65,73,48), [46,66,64]
#> 57: (65,88,14), [63,64,74]
#> 58: (100,73,14), [64,55,46]
#> 59: (4,67,77), [53,84,75]
#> 60: (16,15,68), [65,39,38]
#> 61: (35,16,63), [40,0,38]
#> 62: (90,88,83), [82,76,63]
#> 63: (88,90,14), [55,57,62]
#> 64: (65,14,73), [58,56,57]
#> 65: (34,68,15), [60,108,68]
#> 66: (65,48,4), [75,92,56]
#> 67: (74,44,34), [68,102,54]
#> 68: (44,68,34), [65,67,44]
#> 69: (27,18,74), [81,102,104]
#> 70: (72,59,95), [0,21,34]
#> 71: (90,59,82), [34,35,76]
#> 72: (25,77,74), [54,81,80]
#> 73: (88,70,92), [91,82,74]
#> 74: (88,65,70), [78,73,57]
#> 75: (4,48,67), [51,59,66]
#> 76: (90,83,59), [85,71,62]
#> 77: (9,70,46), [78,89,91]
#> 78: (70,65,46), [79,77,74]
#> 79: (76,46,65), [78,92,114]
#> 80: (53,77,25), [72,101,84]
#> 81: (74,18,25), [101,72,69]
#> 82: (88,92,83), [93,62,73]
#> 83: (69,4,53), [84,96,95]
#> 84: (53,4,77), [59,80,83]
#> 85: (49,59,83), [76,87,90]
#> 86: (49,45,12), [107,90,87]
#> 87: (45,49,83), [85,99,86]
#> 88: (92,79,64), [124,93,94]
#> 89: (46,58,9), [100,77,114]
#> 90: (49,12,59), [0,85,86]
#> 91: (9,92,70), [73,77,94]
#> 92: (76,65,4), [66,98,79]
#> 93: (92,64,83), [106,82,88]
#> 94: (79,92,9), [91,100,88]
#> 95: (31,4,69), [83,112,98]
#> 96: (10,69,53), [83,110,109]
#> 97: (6,31,41), [121,127,134]
#> 98: (31,76,4), [92,95,134]
#> 99: (83,94,45), [107,87,106]
#> 100: (58,79,9), [94,89,120]
#> 101: (53,25,18), [81,122,80]
#> 102: (27,74,34), [67,105,69]
#> 103: (50,15,35), [38,148,113]
#> 104: (18,27,20), [125,116,69]
#> 105: (34,36,27), [118,102,108]
#> 106: (83,64,94), [139,99,93]
#> 107: (94,12,45), [86,99,0]
#> 108: (34,15,36), [113,105,65]
#> 109: (69,10,32), [115,112,96]
#> 110: (53,96,10), [136,96,122]
#> 111: (42,32,17), [115,129,128]
#> 112: (69,32,31), [121,95,109]
#> 113: (50,36,15), [108,103,126]
#> 114: (46,76,58), [135,89,79]
#> 115: (17,32,10), [109,130,111]
#> 116: (18,20,96), [131,122,104]
#> 117: (30,27,57), [118,144,125]
#> 118: (57,27,36), [105,119,117]
#> 119: (24,57,36), [118,126,144]
#> 120: (58,43,79), [124,100,140]
#> 121: (41,31,32), [112,128,97]
#> 122: (18,96,53), [110,101,116]
#> 123: (23,43,38), [140,150,149]
#> 124: (79,43,64), [149,88,120]
#> 125: (30,20,27), [104,117,133]
#> 126: (24,36,50), [113,145,119]
#> 127: (6,41,42), [128,154,97]
#> 128: (42,41,32), [121,111,127]
#> 129: (26,42,17), [111,130,132]
#> 130: (26,17,10), [115,136,129]
#> 131: (20,75,96), [159,116,138]
#> 132: (42,26,99), [155,154,129]
#> 133: (5,20,30), [125,143,142]
#> 134: (6,76,31), [98,97,135]
#> 135: (76,6,58), [152,114,134]
#> 136: (26,10,96), [110,141,130]
#> 137: (40,75,11), [138,160,159]
#> 138: (20,11,75), [137,131,142]
#> 139: (64,23,94), [0,106,149]
#> 140: (38,43,58), [120,150,123]
#> 141: (96,40,26), [166,136,159]
#> 142: (5,11,20), [138,133,161]
#> 143: (30,52,5), [161,133,162]
#> 144: (24,30,57), [117,119,162]
#> 145: (80,24,50), [126,146,157]
#> 146: (50,3,80), [147,145,158]
#> 147: (80,3,60), [158,157,146]
#> 148: (50,35,60), [0,158,103]
#> 149: (23,64,43), [124,123,139]
#> 150: (23,38,58), [140,163,123]
#> 151: (8,6,93), [153,167,152]
#> 152: (8,58,6), [135,151,163]
#> 153: (6,99,93), [165,151,154]
#> 154: (42,99,6), [153,127,132]
#> 155: (26,71,99), [168,132,169]
#> 156: (52,24,60), [157,183,162]
#> 157: (80,60,24), [156,145,147]
#> 158: (50,60,3), [147,146,148]
#> 159: (96,75,40), [137,141,131]
#> 160: (52,40,11), [137,161,170]
#> 161: (52,11,5), [142,143,160]
#> 162: (30,24,52), [156,143,144]
#> 163: (8,23,58), [150,152,177]
#> 164: (7,93,61), [165,176,167]
#> 165: (61,93,99), [153,168,164]
#> 166: (26,40,2), [174,172,141]
#> 167: (7,8,93), [151,164,175]
#> 168: (99,71,61), [178,165,155]
#> 169: (26,55,71), [171,155,172]
#> 170: (91,40,52), [160,183,182]
#> 171: (21,71,55), [169,180,179]
#> 172: (26,2,55), [180,169,166]
#> 173: (91,2,78), [174,182,181]
#> 174: (40,78,2), [173,166,182]
#> 175: (8,7,54), [176,177,167]
#> 176: (61,54,7), [175,164,178]
#> 177: (8,54,23), [0,163,175]
#> 178: (61,71,54), [179,176,168]
#> 179: (21,54,71), [178,171,184]
#> 180: (2,21,55), [171,172,185]
#> 181: (98,2,91), [173,0,185]
#> 182: (91,78,40), [174,170,173]
#> 183: (91,52,60), [156,0,170]
#> 184: (98,54,21), [179,185,0]
#> 185: (98,21,2), [180,181,184]
#> boundary nodes:  1 28 63 35 60 91 98 54 23 94 12 59 95 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.158889487 0.138169820 0.097641909 0.035164742 0.041335406 0.108432525 
#>                                                                         
#> 0.074601549 0.078314909 0.113357109 0.120980566 0.240141117 0.188856299 
#>                                                                         
#> 0.063920571 0.158363224 0.137159441 0.064783465 0.059565280 0.062764894 
#>                                                                         
#> 0.136445138 0.074390696 0.127573288 0.090547282 0.086681294 0.037458304 
#>                                                                         
#> 0.129580857 0.207686935 0.116195176 0.251318188 0.076049043 0.135290994 
#>                                                                         
#> 0.294088941 0.311730218 0.183460376 0.113785698 0.189636129 0.065038109 
#>                                                                         
#> 0.124839045 0.116996996 0.066888894 0.086231568 0.061674942 0.083253785 
#>                                                                         
#> 0.134013600 0.093827867 0.195180412 0.053693077 0.013099408 0.181219195 
#>                                                                         
#> 0.160677296 0.127244313 0.214899188 0.221571473 0.140342159 0.351502797 
#>                                                                         
#> 0.161531485 0.064929851 0.098112054 0.056030731 0.096976971 0.092798722 
#>                                                                         
#> 0.012491042 0.151768595 0.116595292 0.080413247 0.152342760 0.159970846 
#>                                                                         
#> 0.038827431 0.054626407 0.076881032 0.189214971 0.077453327 0.182802704 
#>                                                                         
#> 0.129396491 0.088368384 0.158773281 0.077826160 0.080085852 0.142975701 
#>                                                                         
#> 0.101884774 0.060148623 0.096202237 0.077786280 0.163777298 0.046640042 
#>                                                                         
#> 0.209903097 0.063253710 0.105491779 0.025210184 0.115481212 0.222516234 
#>                                                                         
#> 0.166410567 0.150159358 0.044552616 0.089623550 0.203967172 0.046892251 
#>                                                                         
#> 0.121315021 0.137476477 0.040829143 0.090533078 0.094551792 0.048289414 
#>                                                                         
#> 0.026120052 0.142877187 0.116254785 0.102631210 0.253759403 0.052151263 
#>                                                                         
#> 0.197147774 0.103294692 0.128055388 0.107698247 0.089575768 0.085047850 
#>                                                                         
#> 0.123528085 0.004725712 0.020420494 0.390536007 0.098526742 0.030594938 
#>                                                                         
#> 0.049734433 0.068115808 0.106698647 0.040563438 0.086551764 0.096252548 
#>                                                                         
#> 0.107157303 0.354655407 0.120758184 0.077902316 0.148638682 0.190578996 
#>                                                                         
#> 0.083376791 0.070043424 0.166348407 0.046944116 0.077869520 0.022221963 
#>                                                                         
#> 0.374854976 0.187041112 0.043452774 0.141030745 0.100241421 0.386610617 
#>                                                                         
#> 0.228300680 0.084352692 0.156240499 0.111543029 0.067386909 0.080695649 
#>                                                                         
#> 0.116684269 0.108181237 0.160983830 0.255333039 0.225608871 0.147728355 
#>                                                                         
#> 0.142411173 0.099009028 0.144132948 0.120710763 0.086226742 0.154693780 
#>                                                                         
#> 0.099114992 0.056572007 0.152904011 0.201000007 0.100659060 0.033712428 
#>                                                                         
#> 0.300809049 0.121201164 0.180741927 0.181599520 0.181093849 0.111604458 
#>                                                                         
#> 0.075076270 0.098208243 0.045771473 0.090490179 0.018104562 0.126703954 
#>                                                                         
#> 0.172506248 0.178914073 0.101591651 0.107315430 0.116249446 0.412181896 
#>                                                                         
#> 0.230355258 0.102401853 0.098439128 0.149278019 0.005844530 0.056970004 
#>                                                                         
#> 0.094319187 0.047750898 0.095709433 0.206477842 0.104898569 0.094201721 
#>                                                                         
#> 0.129170899 0.120354739 0.036771514 0.049653514 0.129402402 0.058376810 
#>                                                                         
#> 0.079135171 0.154826121 0.132502622 0.070878607 0.134082325 0.211052993 
#>                                                                         
#> 0.185749575 0.255213065 0.079046975 0.118376058 0.174737184 0.086100493 
#>                                                                         
#> 0.222980078 0.227272424 0.213778151 0.145624755 0.166285873 0.028751742 
#>                                                                         
#> 0.101172334 0.067116322 0.086358221 0.092935910 0.103386101 0.068749928 
#>                                                                         
#> 0.120732080 0.140123335 0.087924734 0.148291265 0.062231620 0.139100963 
#>                                                                         
#> 0.114863809 0.021603970 0.135458834 0.071783328 0.193688304 0.030772521 
#>                                                                         
#> 0.115086771 0.099913167 0.137182804 0.189184773 0.151178342 0.125604816 
#>                                                                         
#> 0.081115190 0.145752457 0.170852738 0.175245906 0.241428982 0.164412669 
#>                                                                         
#> 0.079706976 0.135827608 0.241337875 0.129977279 0.073235457 0.183280175 
#>                                                                         
#> 0.156531300 0.626595596 0.096607375 0.146894474 0.022882240 0.036023771 
#>                                                                         
#> 0.046466289 0.062721408 0.334803443 0.018271451 0.121687705 0.084431121 
#>                                                                         
#> 0.109690959 0.190220209 0.015449892 0.097506761 0.029664275 0.115478494 
#>                                                                         
#> 0.061464339 0.124484688 0.136052728 0.206511228 0.166373878 0.204936435 
#>                         
#> 0.115028776 0.281307109 
#> 
#> $alpha
#> [1] 0.2796098
#> 
#> attr(,"class")
#> [1] "scree"