Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y
numeric vectors

binner
an optional function that bins the x and y vectors prior to triangulation

...
other args

Value

An object of class "scree" that consists of three elements: - `del`: the Delauney-Vornoi tesselation from [alphahull::delvor()] - `weights`: the lengths of each edge in the Delauney triangulation - `alpha`: the radius or `alpha` value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2         x1         y1         x2         y2          mx1         my1
#>    45   76 0.08954979 0.90934941 0.01279798 0.74973864 -0.045598530  0.87607885
#>    76   18 0.01279798 0.74973864 0.05684929 0.77416666 -0.001162272  0.82684641
#>    18   53 0.05684929 0.77416666 0.16902251 0.72701619  0.117201701  0.76073998
#>    53   12 0.16902251 0.72701619 0.16929000 0.87394288  0.149919193  0.80051456
#>    12   45 0.16929000 0.87394288 0.08954979 0.90934941  0.112140362  0.85273033
#>    67   25 0.95018015 0.70991366 0.98329800 0.86705378  0.925020078  0.79727615
#>    25   94 0.98329800 0.86705378 0.84747529 0.88752615  0.909697027  0.83954253
#>    94   10 0.84747529 0.88752615 0.77310852 0.83510123  0.814886852  0.85479558
#>    10   98 0.77310852 0.83510123 0.83603891 0.77197338  0.793249982  0.79224900
#>    98   67 0.83603891 0.77197338 0.95018015 0.70991366  0.921802168  0.79371548
#>    78   93 0.55828061 0.40544079 0.51649272 0.35877566  0.573604590  0.34967564
#>    93   69 0.51649272 0.35877566 0.54281136 0.28865955  0.577671791  0.34174220
#>    69   58 0.54281136 0.28865955 0.71336714 0.24891512  0.638010363  0.31136195
#>    58   56 0.71336714 0.24891512 0.77996648 0.32990144  0.680969554  0.34343461
#>    56    9 0.77996648 0.32990144 0.73305523 0.50976423  0.706012684  0.40666206
#>     9   22 0.73305523 0.50976423 0.67798478 0.56513546  0.647039475  0.47928699
#>    22    7 0.67798478 0.56513546 0.49549658 0.50560304  0.597259328  0.50312582
#>     7   78 0.49549658 0.50560304 0.55828061 0.40544079  0.587851830  0.49373509
#>    90   39 0.05174926 0.15990143 0.17474449 0.04031614  0.136575132  0.12410224
#>    39   96 0.17474449 0.04031614 0.23635621 0.09662820  0.152729665  0.12626393
#>    96   88 0.23635621 0.09662820 0.20792116 0.24390424  0.211016776  0.16811888
#>    88   92 0.20792116 0.24390424 0.14311444 0.31890296  0.159947161  0.26794894
#>    92   90 0.14311444 0.31890296 0.05174926 0.15990143  0.061382578  0.26011677
#>    73   31 0.02087720 0.00000000 0.02407963 0.16085208 -1.804927434  0.11680819
#>    31   99 0.02407963 0.16085208 0.02077116 0.38071851 -0.211424133  0.26726640
#>    99   27 0.02077116 0.38071851 0.04450660 0.40032401  0.021023572  0.40458335
#>    27   59 0.04450660 0.40032401 0.05842017 0.46175393  0.031824739  0.43548702
#>    59   76 0.05842017 0.46175393 0.01279798 0.74973864 -0.086660985  0.58637641
#>    81   51 0.52595083 0.66300066 0.45852371 0.69871753  0.496636360  0.68916380
#>    51   77 0.45852371 0.69871753 0.37342912 0.66654812  0.418503027  0.67594944
#>    77   36 0.37342912 0.66654812 0.42920564 0.57535633  0.419021246  0.63178061
#>    36   35 0.42920564 0.57535633 0.47654800 0.55054806  0.477464632  0.60987385
#>    35   81 0.47654800 0.55054806 0.52595083 0.66300066  0.490867892  0.61133518
#>    22   43 0.67798478 0.56513546 0.64090061 0.61031237  0.588625262  0.52959233
#>    43   35 0.64090061 0.61031237 0.47654800 0.55054806  0.565567481  0.56161138
#>    35   87 0.47654800 0.55054806 0.39742628 0.48264437  0.442780954  0.50984523
#>    87   55 0.39742628 0.48264437 0.49131299 0.45393267  0.448980138  0.48336478
#>    55   78 0.49131299 0.45393267 0.55828061 0.40544079  0.557314081  0.47459332
#>    77   70 0.37342912 0.66654812 0.27402851 0.72075465  0.326341471  0.69844231
#>    70   32 0.27402851 0.72075465 0.22044031 0.66191650  0.253967831  0.68520296
#>    32   57 0.22044031 0.66191650 0.19883291 0.56660089  0.250431172  0.60501084
#>    57   64 0.19883291 0.56660089 0.33174668 0.57967761  0.263342319  0.59293370
#>    64   77 0.33174668 0.57967761 0.37342912 0.66654812  0.331903124  0.63303789
#>    56   79 0.77996648 0.32990144 0.85855326 0.34844030  0.797697371  0.43057497
#>    79   47 0.85855326 0.34844030 0.77023210 0.59192468  0.878507869  0.49343954
#>    47   22 0.77023210 0.59192468 0.67798478 0.56513546  0.729199170  0.56100038
#>    94   61 0.84747529 0.88752615 0.82633629 0.97306127  0.772828034  0.91445764
#>    61   34 0.82633629 0.97306127 0.63563245 0.92653815  0.736154417  0.92860707
#>    34   95 0.63563245 0.92653815 0.71829044 0.82752652  0.732926259  0.92375351
#>    95   98 0.71829044 0.82752652 0.83603891 0.77197338  0.756200833  0.75531573
#>    20   67 0.97825406 0.63254401 0.95018015 0.70991366  0.914868583  0.65332251
#>    98   37 0.83603891 0.77197338 0.70633101 0.68627565  0.791027468  0.69909187
#>    37   47 0.70633101 0.68627565 0.77023210 0.59192468  0.793848151  0.67673375
#>    47   20 0.77023210 0.59192468 0.97825406 0.63254401  0.876410165  0.60113614
#>    83   40 0.25662291 0.85578112 0.21168723 0.90976501  0.213023239  0.86518313
#>    40   24 0.21168723 0.90976501 0.19052548 0.91875723  0.191218067  0.89099061
#>    24   45 0.19052548 0.91875723 0.08954979 0.90934941  0.139921706  0.91529761
#>    53   62 0.16902251 0.72701619 0.26919972 0.81855108  0.193845618  0.80043459
#>    62   83 0.26919972 0.81855108 0.25662291 0.85578112  0.202987032  0.81692288
#>    73   39 0.02087720 0.00000000 0.17474449 0.04031614  0.125944889 -0.08721604
#>    90   31 0.05174926 0.15990143 0.02407963 0.16085208  0.035294684  0.08412637
#>    50   60 0.38576843 0.81834060 0.35042738 0.91760896  0.364571004  0.86671915
#>    60   40 0.35042738 0.91760896 0.21168723 0.90976501  0.278056860  0.96675750
#>    83   21 0.25662291 0.85578112 0.28544155 0.85832746  0.266456945  0.90883604
#>    21   33 0.28544155 0.85832746 0.29649424 0.85622613  0.295121783  0.87912564
#>    33   50 0.29649424 0.85622613 0.38576843 0.81834060  0.345591848  0.84779420
#>   100   66 0.46680712 1.00000000 0.18604045 0.99544852  0.273790313  4.24452143
#>    66   60 0.18604045 0.99544852 0.35042738 0.91760896  0.292478423  1.00772996
#>    60   74 0.35042738 0.91760896 0.46337455 0.94359077  0.405088112  0.93848063
#>    74  100 0.46337455 0.94359077 0.46680712 1.00000000  0.434138049  0.97367890
#>    75   87 0.38662591 0.43575161 0.39742628 0.48264437  0.293202724  0.48195905
#>    87   64 0.39742628 0.48264437 0.33174668 0.57967761  0.293272204  0.48288997
#>    57   59 0.19883291 0.56660089 0.05842017 0.46175393  0.106833716  0.54336270
#>    59   92 0.05842017 0.46175393 0.14311444 0.31890296  0.142255808  0.41492638
#>    92   75 0.14311444 0.31890296 0.38662591 0.43575161  0.261927801  0.38345917
#>    88   52 0.20792116 0.24390424 0.31105561 0.21655297  0.286155640  0.33078383
#>    52   72 0.31105561 0.21655297 0.36791239 0.23509557  0.310280466  0.31537047
#>    72   69 0.36791239 0.23509557 0.54281136 0.28865955  0.450450830  0.27791327
#>    93   55 0.51649272 0.35877566 0.49131299 0.45393267  0.457570622  0.39409408
#>    55   75 0.49131299 0.45393267 0.38662591 0.43575161  0.445893959  0.40497062
#>    76   53 0.01279798 0.74973864 0.16902251 0.72701619  0.081972639  0.67692836
#>    87   36 0.39742628 0.48264437 0.42920564 0.57535633  0.399554419  0.53371747
#>    48   91 0.99357682 0.23361283 1.00000000 0.95294441  4.372078573  0.56313941
#>    91   25 1.00000000 0.95294441 0.98329800 0.86705378  1.919291850  0.72961286
#>    79   48 0.85855326 0.34844030 0.99357682 0.23361283  1.048927816  0.43549870
#>    62   33 0.26919972 0.81855108 0.29649424 0.85622613  0.286661388  0.83462517
#>    37   43 0.70633101 0.68627565 0.64090061 0.61031237  0.702706069  0.62323734
#>    58   26 0.71336714 0.24891512 0.74178495 0.24523025  0.725185033  0.22863316
#>    26   56 0.74178495 0.24523025 0.77996648 0.32990144  0.765745839  0.28536972
#>    66   45 0.18604045 0.99544852 0.08954979 0.90934941  0.136309930  0.95406341
#>    40   66 0.21168723 0.90976501 0.18604045 0.99544852  0.220101714  0.95896369
#>    91   49 1.00000000 0.95294441 0.97317465 0.90619041  1.035583144  0.90145581
#>    49   94 0.97317465 0.90619041 0.84747529 0.88752615  0.914300504  0.87008402
#>    58   30 0.71336714 0.24891512 0.68826657 0.19112443  0.722828472  0.21045934
#>    30   29 0.68826657 0.19112443 0.69561031 0.11554877  0.772649384  0.16117934
#>    29   86 0.69561031 0.11554877 0.92213201 0.10234640  0.811117085  0.14748245
#>    86   38 0.92213201 0.10234640 0.95100127 0.24139118  0.845429393  0.19079122
#>    38   79 0.95100127 0.24139118 0.85855326 0.34844030  0.846415192  0.24451404
#>    99   92 0.02077116 0.38071851 0.14311444 0.31890296  0.075992819  0.33803473
#>    81   65 0.52595083 0.66300066 0.59500998 0.77225285  0.564460306  0.71511103
#>    65   63 0.59500998 0.77225285 0.56851848 0.90280052  0.558473978  0.83280049
#>    63   97 0.56851848 0.90280052 0.54527680 0.90594353  0.547658421  0.83605045
#>    97   85 0.54527680 0.90594353 0.47760959 0.85247777  0.545315625  0.83634109
#>    85   50 0.47760959 0.85247777 0.38576843 0.81834060  0.452290991  0.77998250
#>    50   51 0.38576843 0.81834060 0.45852371 0.69871753  0.450917901  0.77602822
#>    44   82 0.65969981 0.16888057 0.59943038 0.08112133  0.640716652  0.11734252
#>    82   41 0.59943038 0.08112133 0.67978587 0.06650706  0.645230956  0.10473091
#>    41   46 0.67978587 0.06650706 0.76611492 0.03311526  0.733212324  0.07634170
#>    46   86 0.76611492 0.03311526 0.92213201 0.10234640  0.810840468  0.14273634
#>    30   44 0.68826657 0.19112443 0.65969981 0.16888057  0.694549372  0.15359031
#>    69   44 0.54281136 0.28865955 0.65969981 0.16888057  0.621003693  0.24804160
#>    90   99 0.05174926 0.15990143 0.02077116 0.38071851  0.041717764  0.27107560
#>    70   53 0.27402851 0.72075465 0.16902251 0.72701619  0.221008853  0.71522110
#>    53   57 0.16902251 0.72701619 0.19883291 0.56660089  0.119151621  0.63477103
#>    62   50 0.26919972 0.81855108 0.38576843 0.81834060  0.327459916  0.80506778
#>    61   42 0.82633629 0.97306127 0.49657595 0.99722558  0.675524922  1.17713462
#>    42   63 0.49657595 0.99722558 0.56851848 0.90280052  0.577469220  0.98423916
#>    65   95 0.59500998 0.77225285 0.71829044 0.82752652  0.637014862  0.84368368
#>    43   81 0.64090061 0.61031237 0.52595083 0.66300066  0.558250764  0.58173245
#>    95   89 0.71829044 0.82752652 0.67130213 0.68539284  0.686971448  0.75904652
#>    89   43 0.67130213 0.68539284 0.64090061 0.61031237  0.690120967  0.63407742
#>    86   48 0.92213201 0.10234640 0.99357682 0.23361283  0.959430878  0.16712159
#>    73   46 0.02087720 0.00000000 0.76611492 0.03311526  0.453689636 -1.33806003
#>    46   84 0.76611492 0.03311526 0.28560107 0.07283984  0.461087735 -0.73049233
#>    84   96 0.28560107 0.07283984 0.23635621 0.09662820  0.236673682  0.03441975
#>    82   46 0.59943038 0.08112133 0.76611492 0.03311526  0.557402995 -0.37818468
#>    74   97 0.46337455 0.94359077 0.54527680 0.90594353  0.515263934  0.94856352
#>    61  100 0.82633629 0.97306127 0.46680712 1.00000000  1.327344429 10.07224788
#>    89   81 0.67130213 0.68539284 0.52595083 0.66300066  0.599131135  0.67092095
#>    69   68 0.54281136 0.28865955 0.54066621 0.21147062  0.620824990  0.24786721
#>    68   82 0.54066621 0.21147062 0.59943038 0.08112133  0.587259102  0.15405496
#>    82   84 0.59943038 0.08112133 0.28560107 0.07283984  0.463161727 -0.70540504
#>    91   61 1.00000000 0.95294441 0.82633629 0.97306127  0.914111954  0.97115052
#>    62   80 0.26919972 0.81855108 0.38134973 0.73716598  0.327416114  0.78080940
#>    80   51 0.38134973 0.73716598 0.45852371 0.69871753  0.447987833  0.77424614
#>    85   71 0.47760959 0.85247777 0.44388853 0.86631570  0.440822162  0.81083774
#>    71   60 0.44388853 0.86631570 0.35042738 0.91760896  0.387853047  0.87500791
#>    80   77 0.38134973 0.73716598 0.37342912 0.66654812  0.410095641  0.69818867
#>    84   72 0.28560107 0.07283984 0.36791239 0.23509557  0.370087153  0.13198645
#>    70   62 0.27402851 0.72075465 0.26919972 0.81855108  0.224114783  0.76730754
#>    76   57 0.01279798 0.74973864 0.19883291 0.56660089  0.074637805  0.62649891
#>    71   74 0.44388853 0.86631570 0.46337455 0.94359077  0.410286276  0.91588337
#>    70   80 0.27402851 0.72075465 0.38134973 0.73716598  0.321092916  0.77209593
#>    85   74 0.47760959 0.85247777 0.46337455 0.94359077  0.477023275  0.89905467
#>     1   76 0.07411634 0.80508343 0.01279798 0.74973864 -0.045598530  0.87607885
#>     1   18 0.07411634 0.80508343 0.05684929 0.77416666 -0.001162272  0.82684641
#>     1   53 0.07411634 0.80508343 0.16902251 0.72701619  0.117201701  0.76073998
#>     1   12 0.07411634 0.80508343 0.16929000 0.87394288  0.149919193  0.80051456
#>     1   45 0.07411634 0.80508343 0.08954979 0.90934941  0.112140362  0.85273033
#>     2   25 0.83556524 0.81349545 0.98329800 0.86705378  0.925020078  0.79727615
#>     2   94 0.83556524 0.81349545 0.84747529 0.88752615  0.909697027  0.83954253
#>     2   10 0.83556524 0.81349545 0.77310852 0.83510123  0.814886852  0.85479558
#>     2   98 0.83556524 0.81349545 0.83603891 0.77197338  0.793249982  0.79224900
#>     2   67 0.83556524 0.81349545 0.95018015 0.70991366  0.921802168  0.79371548
#>     3   93 0.59955503 0.40135882 0.51649272 0.35877566  0.573604590  0.34967564
#>     3   69 0.59955503 0.40135882 0.54281136 0.28865955  0.577671791  0.34174220
#>     3   58 0.59955503 0.40135882 0.71336714 0.24891512  0.638010363  0.31136195
#>     3   56 0.59955503 0.40135882 0.77996648 0.32990144  0.680969554  0.34343461
#>     3    9 0.59955503 0.40135882 0.73305523 0.50976423  0.706012684  0.40666206
#>     3   22 0.59955503 0.40135882 0.67798478 0.56513546  0.647039475  0.47928699
#>     3    7 0.59955503 0.40135882 0.49549658 0.50560304  0.597259328  0.50312582
#>     3   78 0.59955503 0.40135882 0.55828061 0.40544079  0.587851830  0.49373509
#>     4   39 0.15137273 0.21497601 0.17474449 0.04031614  0.136575132  0.12410224
#>     4   96 0.15137273 0.21497601 0.23635621 0.09662820  0.152729665  0.12626393
#>     4   88 0.15137273 0.21497601 0.20792116 0.24390424  0.211016776  0.16811888
#>     4   92 0.15137273 0.21497601 0.14311444 0.31890296  0.159947161  0.26794894
#>     4   90 0.15137273 0.21497601 0.05174926 0.15990143  0.061382578  0.26011677
#>     5   31 0.00000000 0.41587955 0.02407963 0.16085208 -1.804927434  0.11680819
#>     5   99 0.00000000 0.41587955 0.02077116 0.38071851 -0.211424133  0.26726640
#>     5   27 0.00000000 0.41587955 0.04450660 0.40032401  0.021023572  0.40458335
#>     5   59 0.00000000 0.41587955 0.05842017 0.46175393  0.031824739  0.43548702
#>     5   76 0.00000000 0.41587955 0.01279798 0.74973864 -0.086660985  0.58637641
#>     6   51 0.46378510 0.66760823 0.45852371 0.69871753  0.496636360  0.68916380
#>     6   77 0.46378510 0.66760823 0.37342912 0.66654812  0.418503027  0.67594944
#>     6   36 0.46378510 0.66760823 0.42920564 0.57535633  0.419021246  0.63178061
#>     6   35 0.46378510 0.66760823 0.47654800 0.55054806  0.477464632  0.60987385
#>     6   81 0.46378510 0.66760823 0.52595083 0.66300066  0.490867892  0.61133518
#>     7   43 0.49549658 0.50560304 0.64090061 0.61031237  0.588625262  0.52959233
#>     7   35 0.49549658 0.50560304 0.47654800 0.55054806  0.565567481  0.56161138
#>     7   87 0.49549658 0.50560304 0.39742628 0.48264437  0.442780954  0.50984523
#>     7   55 0.49549658 0.50560304 0.49131299 0.45393267  0.448980138  0.48336478
#>     8   70 0.28531520 0.65905536 0.27402851 0.72075465  0.326341471  0.69844231
#>     8   32 0.28531520 0.65905536 0.22044031 0.66191650  0.253967831  0.68520296
#>     8   57 0.28531520 0.65905536 0.19883291 0.56660089  0.250431172  0.60501084
#>     8   64 0.28531520 0.65905536 0.33174668 0.57967761  0.263342319  0.59293370
#>     8   77 0.28531520 0.65905536 0.37342912 0.66654812  0.331903124  0.63303789
#>     9   79 0.73305523 0.50976423 0.85855326 0.34844030  0.797697371  0.43057497
#>     9   47 0.73305523 0.50976423 0.77023210 0.59192468  0.878507869  0.49343954
#>    10   61 0.77310852 0.83510123 0.82633629 0.97306127  0.772828034  0.91445764
#>    10   34 0.77310852 0.83510123 0.63563245 0.92653815  0.736154417  0.92860707
#>    10   95 0.77310852 0.83510123 0.71829044 0.82752652  0.732926259  0.92375351
#>    11   67 0.87625318 0.70771291 0.95018015 0.70991366  0.914868583  0.65332251
#>    11   98 0.87625318 0.70771291 0.83603891 0.77197338  0.911266066  0.77433731
#>    11   37 0.87625318 0.70771291 0.70633101 0.68627565  0.791027468  0.69909187
#>    11   47 0.87625318 0.70771291 0.77023210 0.59192468  0.793848151  0.67673375
#>    11   20 0.87625318 0.70771291 0.97825406 0.63254401  0.876410165  0.60113614
#>    12   40 0.16929000 0.87394288 0.21168723 0.90976501  0.213023239  0.86518313
#>    12   24 0.16929000 0.87394288 0.19052548 0.91875723  0.191218067  0.89099061
#>    12   62 0.16929000 0.87394288 0.26919972 0.81855108  0.193845618  0.80043459
#>    12   83 0.16929000 0.87394288 0.25662291 0.85578112  0.202987032  0.81692288
#>    13   39 0.02712207 0.00701722 0.17474449 0.04031614  0.125944889 -0.08721604
#>    13   90 0.02712207 0.00701722 0.05174926 0.15990143  0.089256214  0.07543404
#>    13   31 0.02712207 0.00701722 0.02407963 0.16085208  0.035294684  0.08412637
#>    13   73 0.02712207 0.00701722 0.02087720 0.00000000 -0.064373773  0.08215520
#>    14   60 0.31625328 0.88805487 0.35042738 0.91760896  0.364571004  0.86671915
#>    14   40 0.31625328 0.88805487 0.21168723 0.90976501  0.278056860  0.96675750
#>    14   83 0.31625328 0.88805487 0.25662291 0.85578112  0.266149187  0.90940467
#>    14   21 0.31625328 0.88805487 0.28544155 0.85832746  0.266456945  0.90883604
#>    14   33 0.31625328 0.88805487 0.29649424 0.85622613  0.295121783  0.87912564
#>    14   50 0.31625328 0.88805487 0.38576843 0.81834060  0.345591848  0.84779420
#>    15   66 0.39905112 0.99667837 0.18604045 0.99544852  0.273790313  4.24452143
#>    15   60 0.39905112 0.99667837 0.35042738 0.91760896  0.292478423  1.00772996
#>    15   74 0.39905112 0.99667837 0.46337455 0.94359077  0.405088112  0.93848063
#>    15  100 0.39905112 0.99667837 0.46680712 1.00000000  0.434138049  0.97367890
#>    16   87 0.19023559 0.49810795 0.39742628 0.48264437  0.293202724  0.48195905
#>    16   64 0.19023559 0.49810795 0.33174668 0.57967761  0.293272204  0.48288997
#>    16   57 0.19023559 0.49810795 0.19883291 0.56660089  0.270237297  0.52285208
#>    16   59 0.19023559 0.49810795 0.05842017 0.46175393  0.106833716  0.54336270
#>    16   92 0.19023559 0.49810795 0.14311444 0.31890296  0.142255808  0.41492638
#>    16   75 0.19023559 0.49810795 0.38662591 0.43575161  0.261927801  0.38345917
#>    17   92 0.40027363 0.35619608 0.14311444 0.31890296  0.266382547  0.37417552
#>    17   88 0.40027363 0.35619608 0.20792116 0.24390424  0.268219636  0.36150765
#>    17   52 0.40027363 0.35619608 0.31105561 0.21655297  0.286155640  0.33078383
#>    17   72 0.40027363 0.35619608 0.36791239 0.23509557  0.310280466  0.31537047
#>    17   69 0.40027363 0.35619608 0.54281136 0.28865955  0.450450830  0.27791327
#>    17   93 0.40027363 0.35619608 0.51649272 0.35877566  0.459715362  0.29746632
#>    17   55 0.40027363 0.35619608 0.49131299 0.45393267  0.457570622  0.39409408
#>    17   75 0.40027363 0.35619608 0.38662591 0.43575161  0.445893959  0.40497062
#>    19   87 0.38528195 0.58280195 0.39742628 0.48264437  0.361556625  0.52911015
#>    19   36 0.38528195 0.58280195 0.42920564 0.57535633  0.399554419  0.53371747
#>    19   77 0.38528195 0.58280195 0.37342912 0.66654812  0.415848402  0.62983997
#>    19   64 0.38528195 0.58280195 0.33174668 0.57967761  0.356170922  0.62139365
#>    20   91 0.97825406 0.63254401 1.00000000 0.95294441  4.372078573  0.56313941
#>    20   25 0.97825406 0.63254401 0.98329800 0.86705378  1.919291850  0.72961286
#>    20   79 0.97825406 0.63254401 0.85855326 0.34844030  0.896187271  0.49985254
#>    20   48 0.97825406 0.63254401 0.99357682 0.23361283  1.048927816  0.43549870
#>    21   62 0.28544155 0.85832746 0.26919972 0.81855108  0.272503194  0.84040637
#>    22   37 0.67798478 0.56513546 0.70633101 0.68627565  0.711738979  0.62112368
#>    23   26 0.73551829 0.24971027 0.74178495 0.24523025  0.725185033  0.22863316
#>    23   56 0.73551829 0.24971027 0.77996648 0.32990144  0.765745839  0.28536972
#>    23   58 0.73551829 0.24971027 0.71336714 0.24891512  0.722283647  0.30945985
#>    24   66 0.19052548 0.91875723 0.18604045 0.99544852  0.220101714  0.95896369
#>    25   49 0.98329800 0.86705378 0.97317465 0.90619041  1.035583144  0.90145581
#>    26   30 0.74178495 0.24523025 0.68826657 0.19112443  0.722828472  0.21045934
#>    26   29 0.74178495 0.24523025 0.69561031 0.11554877  0.772649384  0.16117934
#>    26   86 0.74178495 0.24523025 0.92213201 0.10234640  0.811117085  0.14748245
#>    26   38 0.74178495 0.24523025 0.95100127 0.24139118  0.845429393  0.19079122
#>    26   79 0.74178495 0.24523025 0.85855326 0.34844030  0.846415192  0.24451404
#>    27   92 0.04450660 0.40032401 0.14311444 0.31890296  0.075992819  0.33803473
#>    28   65 0.52827022 0.76885810 0.59500998 0.77225285  0.564460306  0.71511103
#>    28   63 0.52827022 0.76885810 0.56851848 0.90280052  0.558473978  0.83280049
#>    28   97 0.52827022 0.76885810 0.54527680 0.90594353  0.547658421  0.83605045
#>    28   85 0.52827022 0.76885810 0.47760959 0.85247777  0.545315625  0.83634109
#>    28   50 0.52827022 0.76885810 0.38576843 0.81834060  0.452290991  0.77998250
#>    28   51 0.52827022 0.76885810 0.45852371 0.69871753  0.450917901  0.77602822
#>    28   81 0.52827022 0.76885810 0.52595083 0.66300066  0.511001346  0.71628234
#>    29   82 0.69561031 0.11554877 0.59943038 0.08112133  0.640716652  0.11734252
#>    29   41 0.69561031 0.11554877 0.67978587 0.06650706  0.645230956  0.10473091
#>    29   46 0.69561031 0.11554877 0.76611492 0.03311526  0.733212324  0.07634170
#>    29   44 0.69561031 0.11554877 0.65969981 0.16888057  0.694549372  0.15359031
#>    30   69 0.68826657 0.19112443 0.54281136 0.28865955  0.624453908  0.25318695
#>    32   53 0.22044031 0.66191650 0.16902251 0.72701619  0.221008853  0.71522110
#>    34   42 0.63563245 0.92653815 0.49657595 0.99722558  0.675524922  1.17713462
#>    34   63 0.63563245 0.92653815 0.56851848 0.90280052  0.577469220  0.98423916
#>    34   65 0.63563245 0.92653815 0.59500998 0.77225285  0.626172811  0.84653834
#>    37   95 0.70633101 0.68627565 0.71829044 0.82752652  0.755234178  0.75326685
#>    37   89 0.70633101 0.68627565 0.67130213 0.68539284  0.686971448  0.75904652
#>    38   48 0.95100127 0.24139118 0.99357682 0.23361283  0.959430878  0.16712159
#>    39   46 0.17474449 0.04031614 0.76611492 0.03311526  0.453689636 -1.33806003
#>    39   84 0.17474449 0.04031614 0.28560107 0.07283984  0.461087735 -0.73049233
#>    42   74 0.49657595 0.99722558 0.46337455 0.94359077  0.479112661  0.97094214
#>    42   97 0.49657595 0.99722558 0.54527680 0.90594353  0.515263934  0.94856352
#>    42  100 0.49657595 0.99722558 0.46680712 1.00000000  1.327344429 10.07224788
#>    44   68 0.65969981 0.16888057 0.54066621 0.21147062  0.620824990  0.24786721
#>    49   61 0.97317465 0.90619041 0.82633629 0.97306127  0.914111954  0.97115052
#>    50   80 0.38576843 0.81834060 0.38134973 0.73716598  0.327416114  0.78080940
#>    50   71 0.38576843 0.81834060 0.44388853 0.86631570  0.440822162  0.81083774
#>    52   84 0.31105561 0.21655297 0.28560107 0.07283984  0.290605050  0.14606436
#>    52   96 0.31105561 0.21655297 0.23635621 0.09662820  0.244743805  0.17463066
#>    54   69 0.52964550 0.21362291 0.54281136 0.28865955  0.542478984  0.25004451
#>    54   72 0.52964550 0.21362291 0.36791239 0.23509557  0.454244498  0.26552606
#>    54   84 0.52964550 0.21362291 0.28560107 0.07283984  0.432323293  0.10041444
#>    54   82 0.52964550 0.21362291 0.59943038 0.08112133  0.442356288  0.08302246
#>    54   68 0.52964550 0.21362291 0.54066621 0.21147062  0.517599660  0.12265104
#>    65   89 0.59500998 0.77225285 0.67130213 0.68539284  0.595329433  0.69559842
#>     5   73 0.00000000 0.41587955 0.02087720 0.00000000 -1.804927434  0.11680819
#>           mx2         my2 bp1 bp2
#>  -5.134183027  3.32301947   0   1
#>   0.081972639  0.67692836   0   0
#>   0.081972639  0.67692836   0   0
#>   0.193845618  0.80043459   0   0
#>   0.139921706  0.91529761   0   0
#>   1.169560131  0.74573844   0   0
#>   0.914300504  0.87008402   0   0
#>   0.772828034  0.91445764   0   0
#>   0.756200833  0.75531573   0   0
#>   0.911266066  0.77433731   0   0
#>   0.508849239  0.40766304   0   0
#>   0.459715362  0.29746632   0   0
#>   0.624453908  0.25318695   0   0
#>   0.722283647  0.30945985   0   0
#>   0.797697371  0.43057497   0   0
#>   0.729199170  0.56100038   0   0
#>   0.588625262  0.52959233   0   0
#>   0.557314081  0.47459332   0   0
#>   0.089256214  0.07543404   0   0
#>   0.236673682  0.03441975   0   0
#>   0.244743805  0.17463066   0   0
#>   0.268219636  0.36150765   0   0
#>   0.042194646  0.27114250   0   0
#>  -0.064373773  0.08215520   0   0
#>   0.041717764  0.27107560   0   0
#>   0.075992819  0.33803473   0   0
#>   0.136823123  0.41170542   0   0
#>   0.057644889  0.60923717   0   0
#>   0.511001346  0.71628234   0   0
#>   0.410095641  0.69818867   0   0
#>   0.415848402  0.62983997   0   0
#>   0.431770594  0.52267456   0   0
#>   0.558250764  0.58173245   0   0
#>   0.702706069  0.62323734   0   0
#>   0.558250764  0.58173245   0   0
#>   0.431770594  0.52267456   0   0
#>   0.438328346  0.44853363   0   0
#>   0.508849239  0.40766304   0   0
#>   0.331038708  0.70705581   0   0
#>   0.221008853  0.71522110   0   0
#>   0.119151621  0.63477103   0   0
#>   0.270237297  0.52285208   0   0
#>   0.356170922  0.62139365   0   0
#>   0.839832944  0.25196097   0   0
#>   0.896187271  0.49985254   0   0
#>   0.711738979  0.62112368   0   0
#>   0.902937284  0.94661263   0   0
#>   0.675524922  1.17713462   0   0
#>   0.637014862  0.84368368   0   0
#>   0.755234178  0.75326685   0   0
#>   1.169560131  0.74573844   0   0
#>   0.755234178  0.75326685   0   0
#>   0.711738979  0.62112368   0   0
#>   0.896187271  0.49985254   0   0
#>   0.266149187  0.90940467   0   0
#>   0.220101714  0.95896369   0   0
#>   0.136309930  0.95406341   0   0
#>   0.224114783  0.76730754   0   0
#>   0.272503194  0.84040637   0   0
#>   0.453689636 -1.33806003   0   0
#>   0.041717764  0.27107560   0   0
#>   0.387853047  0.87500791   0   0
#>   0.277525732  0.97615184   0   0
#>   0.272503194  0.84040637   0   0
#>   0.286661388  0.83462517   0   0
#>   0.327459916  0.80506778   0   0
#>   0.216067543  7.80526196   0   1
#>   0.277525732  0.97615184   0   0
#>   0.410286276  0.91588337   0   0
#>   0.479112661  0.97094214   0   0
#>   0.438328346  0.44853363   0   0
#>   0.361556625  0.52911015   0   0
#>   0.057644889  0.60923717   0   0
#>   0.136823123  0.41170542   0   0
#>   0.266382547  0.37417552   0   0
#>   0.244743805  0.17463066   0   0
#>   0.370087153  0.13198645   0   0
#>   0.454244498  0.26552606   0   0
#>   0.508849239  0.40766304   0   0
#>   0.438328346  0.44853363   0   0
#>   0.074637805  0.62649891   0   0
#>   0.431770594  0.52267456   0   0
#>   5.762147216  0.55072698   0   1
#>   1.035583144  0.90145581   0   0
#>   0.997390424  0.37489680   0   0
#>   0.327459916  0.80506778   0   0
#>   0.690120967  0.63407742   0   0
#>   0.722828472  0.21045934   0   0
#>   0.839832944  0.25196097   0   0
#>  -5.012079897  6.72382657   0   1
#>   0.277525732  0.97615184   0   0
#>   0.914111954  0.97115052   0   0
#>   0.902937284  0.94661263   0   0
#>   0.624453908  0.25318695   0   0
#>   0.694549372  0.15359031   0   0
#>   0.810840468  0.14273634   0   0
#>   0.959430878  0.16712159   0   0
#>   0.997390424  0.37489680   0   0
#>   0.042194646  0.27114250   0   0
#>   0.595329433  0.69559842   0   0
#>   0.626172811  0.84653834   0   0
#>   0.566575804  0.97593946   0   0
#>   0.493704637  0.90166088   0   0
#>   0.440822162  0.81083774   0   0
#>   0.447987833  0.77424614   0   0
#>   0.587259102  0.15405496   0   0
#>   0.557402995 -0.37818468   0   0
#>   0.557402995 -0.37818468   0   0
#>   3.187117992 -5.21236686   0   1
#>   0.621003693  0.24804160   0   0
#>   0.620824990  0.24786721   0   0
#>   0.042194646  0.27114250   0   0
#>   0.224114783  0.76730754   0   0
#>   0.074637805  0.62649891   0   0
#>   0.327416114  0.78080940   0   0
#>   1.327344429 10.07224788   0   0
#>   0.566575804  0.97593946   0   0
#>   0.672872788  0.76370743   0   0
#>   0.599131135  0.67092095   0   0
#>   0.672872788  0.76370743   0   0
#>   0.599131135  0.67092095   0   0
#>   6.836531307 -3.03162727   0   1
#>   0.513198530 -2.67726940   0   1
#>   0.463161727 -0.70540504   0   0
#>   0.290605050  0.14606436   0   0
#>   0.463161727 -0.70540504   0   0
#>   0.493704637  0.90166088   0   0
#>   1.534586257 12.83813521   0   1
#>   0.595329433  0.69559842   0   0
#>   0.542478984  0.25004451   0   0
#>   0.517599660  0.12265104   0   0
#>   0.442356288  0.08302246   0   0
#>   1.572304935  6.65316206   0   1
#>   0.321092916  0.77209593   0   0
#>   0.410095641  0.69818867   0   0
#>   0.477023275  0.89905467   0   0
#>   0.410286276  0.91588337   0   0
#>   0.331038708  0.70705581   0   0
#>   0.432323293  0.10041444   0   0
#>   0.321092916  0.77209593   0   0
#>   0.057644889  0.60923717   0   0
#>   0.477023275  0.89905467   0   0
#>   0.331038708  0.70705581   0   0
#>   0.493704637  0.90166088   0   0
#>  -0.001162272  0.82684641   0   0
#>   0.117201701  0.76073998   0   0
#>   0.149919193  0.80051456   0   0
#>   0.112140362  0.85273033   0   0
#>  -0.045598530  0.87607885   0   0
#>   0.909697027  0.83954253   0   0
#>   0.814886852  0.85479558   0   0
#>   0.793249982  0.79224900   0   0
#>   0.921802168  0.79371548   0   0
#>   0.925020078  0.79727615   0   0
#>   0.577671791  0.34174220   0   0
#>   0.638010363  0.31136195   0   0
#>   0.680969554  0.34343461   0   0
#>   0.706012684  0.40666206   0   0
#>   0.647039475  0.47928699   0   0
#>   0.597259328  0.50312582   0   0
#>   0.587851830  0.49373509   0   0
#>   0.573604590  0.34967564   0   0
#>   0.152729665  0.12626393   0   0
#>   0.211016776  0.16811888   0   0
#>   0.159947161  0.26794894   0   0
#>   0.061382578  0.26011677   0   0
#>   0.136575132  0.12410224   0   0
#>  -0.211424133  0.26726640   0   0
#>   0.021023572  0.40458335   0   0
#>   0.031824739  0.43548702   0   0
#>  -0.086660985  0.58637641   0   0
#>  -3.077541662  0.70102730   0   1
#>   0.418503027  0.67594944   0   0
#>   0.419021246  0.63178061   0   0
#>   0.477464632  0.60987385   0   0
#>   0.490867892  0.61133518   0   0
#>   0.496636360  0.68916380   0   0
#>   0.565567481  0.56161138   0   0
#>   0.442780954  0.50984523   0   0
#>   0.448980138  0.48336478   0   0
#>   0.557314081  0.47459332   0   0
#>   0.253967831  0.68520296   0   0
#>   0.250431172  0.60501084   0   0
#>   0.263342319  0.59293370   0   0
#>   0.331903124  0.63303789   0   0
#>   0.326341471  0.69844231   0   0
#>   0.878507869  0.49343954   0   0
#>   0.729199170  0.56100038   0   0
#>   0.736154417  0.92860707   0   0
#>   0.732926259  0.92375351   0   0
#>   0.756200833  0.75531573   0   0
#>   0.911266066  0.77433731   0   0
#>   0.791027468  0.69909187   0   0
#>   0.793848151  0.67673375   0   0
#>   0.876410165  0.60113614   0   0
#>   0.914868583  0.65332251   0   0
#>   0.191218067  0.89099061   0   0
#>   0.139921706  0.91529761   0   0
#>   0.202987032  0.81692288   0   0
#>   0.213023239  0.86518313   0   0
#>   0.089256214  0.07543404   0   0
#>   0.035294684  0.08412637   0   0
#>  -0.064373773  0.08215520   0   0
#>   0.125944889 -0.08721604   0   0
#>   0.278056860  0.96675750   0   0
#>   0.266149187  0.90940467   0   0
#>   0.266456945  0.90883604   0   0
#>   0.295121783  0.87912564   0   0
#>   0.345591848  0.84779420   0   0
#>   0.364571004  0.86671915   0   0
#>   0.292478423  1.00772996   0   0
#>   0.405088112  0.93848063   0   0
#>   0.434138049  0.97367890   0   0
#>   0.273790313  4.24452143   0   0
#>   0.293272204  0.48288997   0   0
#>   0.270237297  0.52285208   0   0
#>   0.106833716  0.54336270   0   0
#>   0.142255808  0.41492638   0   0
#>   0.261927801  0.38345917   0   0
#>   0.293202724  0.48195905   0   0
#>   0.268219636  0.36150765   0   0
#>   0.286155640  0.33078383   0   0
#>   0.310280466  0.31537047   0   0
#>   0.450450830  0.27791327   0   0
#>   0.459715362  0.29746632   0   0
#>   0.457570622  0.39409408   0   0
#>   0.445893959  0.40497062   0   0
#>   0.266382547  0.37417552   0   0
#>   0.399554419  0.53371747   0   0
#>   0.415848402  0.62983997   0   0
#>   0.356170922  0.62139365   0   0
#>   0.361556625  0.52911015   0   0
#>   1.919291850  0.72961286   0   0
#>   1.169560131  0.74573844   0   0
#>   1.048927816  0.43549870   0   0
#>   4.372078573  0.56313941   0   0
#>   0.286661388  0.83462517   0   0
#>   0.702706069  0.62323734   0   0
#>   0.765745839  0.28536972   0   0
#>   0.722283647  0.30945985   0   0
#>   0.725185033  0.22863316   0   0
#>   0.136309930  0.95406341   0   0
#>   0.914300504  0.87008402   0   0
#>   0.772649384  0.16117934   0   0
#>   0.811117085  0.14748245   0   0
#>   0.845429393  0.19079122   0   0
#>   0.846415192  0.24451404   0   0
#>   0.839832944  0.25196097   0   0
#>   0.136823123  0.41170542   0   0
#>   0.558473978  0.83280049   0   0
#>   0.547658421  0.83605045   0   0
#>   0.545315625  0.83634109   0   0
#>   0.452290991  0.77998250   0   0
#>   0.450917901  0.77602822   0   0
#>   0.511001346  0.71628234   0   0
#>   0.564460306  0.71511103   0   0
#>   0.645230956  0.10473091   0   0
#>   0.733212324  0.07634170   0   0
#>   0.810840468  0.14273634   0   0
#>   0.640716652  0.11734252   0   0
#>   0.621003693  0.24804160   0   0
#>   0.119151621  0.63477103   0   0
#>   0.577469220  0.98423916   0   0
#>   0.626172811  0.84653834   0   0
#>   0.637014862  0.84368368   0   0
#>   0.686971448  0.75904652   0   0
#>   0.690120967  0.63407742   0   0
#>   0.997390424  0.37489680   0   0
#>   0.461087735 -0.73049233   0   0
#>   0.236673682  0.03441975   0   0
#>   0.515263934  0.94856352   0   0
#>   0.566575804  0.97593946   0   0
#>   0.479112661  0.97094214   0   0
#>   0.587259102  0.15405496   0   0
#>   0.902937284  0.94661263   0   0
#>   0.447987833  0.77424614   0   0
#>   0.387853047  0.87500791   0   0
#>   0.370087153  0.13198645   0   0
#>   0.290605050  0.14606436   0   0
#>   0.454244498  0.26552606   0   0
#>   0.432323293  0.10041444   0   0
#>   0.442356288  0.08302246   0   0
#>   0.517599660  0.12265104   0   0
#>   0.542478984  0.25004451   0   0
#>   0.672872788  0.76370743   0   0
#>  -4.203425452 -0.00359668   0   1
#> 
#> $x
#>              [,1]       [,2]
#>   [1,] 0.07411634 0.80508343
#>   [2,] 0.83556524 0.81349545
#>   [3,] 0.59955503 0.40135882
#>   [4,] 0.15137273 0.21497601
#>   [5,] 0.00000000 0.41587955
#>   [6,] 0.46378510 0.66760823
#>   [7,] 0.49549658 0.50560304
#>   [8,] 0.28531520 0.65905536
#>   [9,] 0.73305523 0.50976423
#>  [10,] 0.77310852 0.83510123
#>  [11,] 0.87625318 0.70771291
#>  [12,] 0.16929000 0.87394288
#>  [13,] 0.02712207 0.00701722
#>  [14,] 0.31625328 0.88805487
#>  [15,] 0.39905112 0.99667837
#>  [16,] 0.19023559 0.49810795
#>  [17,] 0.40027363 0.35619608
#>  [18,] 0.05684929 0.77416666
#>  [19,] 0.38528195 0.58280195
#>  [20,] 0.97825406 0.63254401
#>  [21,] 0.28544155 0.85832746
#>  [22,] 0.67798478 0.56513546
#>  [23,] 0.73551829 0.24971027
#>  [24,] 0.19052548 0.91875723
#>  [25,] 0.98329800 0.86705378
#>  [26,] 0.74178495 0.24523025
#>  [27,] 0.04450660 0.40032401
#>  [28,] 0.52827022 0.76885810
#>  [29,] 0.69561031 0.11554877
#>  [30,] 0.68826657 0.19112443
#>  [31,] 0.02407963 0.16085208
#>  [32,] 0.22044031 0.66191650
#>  [33,] 0.29649424 0.85622613
#>  [34,] 0.63563245 0.92653815
#>  [35,] 0.47654800 0.55054806
#>  [36,] 0.42920564 0.57535633
#>  [37,] 0.70633101 0.68627565
#>  [38,] 0.95100127 0.24139118
#>  [39,] 0.17474449 0.04031614
#>  [40,] 0.21168723 0.90976501
#>  [41,] 0.67978587 0.06650706
#>  [42,] 0.49657595 0.99722558
#>  [43,] 0.64090061 0.61031237
#>  [44,] 0.65969981 0.16888057
#>  [45,] 0.08954979 0.90934941
#>  [46,] 0.76611492 0.03311526
#>  [47,] 0.77023210 0.59192468
#>  [48,] 0.99357682 0.23361283
#>  [49,] 0.97317465 0.90619041
#>  [50,] 0.38576843 0.81834060
#>  [51,] 0.45852371 0.69871753
#>  [52,] 0.31105561 0.21655297
#>  [53,] 0.16902251 0.72701619
#>  [54,] 0.52964550 0.21362291
#>  [55,] 0.49131299 0.45393267
#>  [56,] 0.77996648 0.32990144
#>  [57,] 0.19883291 0.56660089
#>  [58,] 0.71336714 0.24891512
#>  [59,] 0.05842017 0.46175393
#>  [60,] 0.35042738 0.91760896
#>  [61,] 0.82633629 0.97306127
#>  [62,] 0.26919972 0.81855108
#>  [63,] 0.56851848 0.90280052
#>  [64,] 0.33174668 0.57967761
#>  [65,] 0.59500998 0.77225285
#>  [66,] 0.18604045 0.99544852
#>  [67,] 0.95018015 0.70991366
#>  [68,] 0.54066621 0.21147062
#>  [69,] 0.54281136 0.28865955
#>  [70,] 0.27402851 0.72075465
#>  [71,] 0.44388853 0.86631570
#>  [72,] 0.36791239 0.23509557
#>  [73,] 0.02087720 0.00000000
#>  [74,] 0.46337455 0.94359077
#>  [75,] 0.38662591 0.43575161
#>  [76,] 0.01279798 0.74973864
#>  [77,] 0.37342912 0.66654812
#>  [78,] 0.55828061 0.40544079
#>  [79,] 0.85855326 0.34844030
#>  [80,] 0.38134973 0.73716598
#>  [81,] 0.52595083 0.66300066
#>  [82,] 0.59943038 0.08112133
#>  [83,] 0.25662291 0.85578112
#>  [84,] 0.28560107 0.07283984
#>  [85,] 0.47760959 0.85247777
#>  [86,] 0.92213201 0.10234640
#>  [87,] 0.39742628 0.48264437
#>  [88,] 0.20792116 0.24390424
#>  [89,] 0.67130213 0.68539284
#>  [90,] 0.05174926 0.15990143
#>  [91,] 1.00000000 0.95294441
#>  [92,] 0.14311444 0.31890296
#>  [93,] 0.51649272 0.35877566
#>  [94,] 0.84747529 0.88752615
#>  [95,] 0.71829044 0.82752652
#>  [96,] 0.23635621 0.09662820
#>  [97,] 0.54527680 0.90594353
#>  [98,] 0.83603891 0.77197338
#>  [99,] 0.02077116 0.38071851
#> [100,] 0.46680712 1.00000000
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.07411634,0.8050834) [5]: 12 18 45 53 76 
#> 2: (0.8355652,0.8134955) [5]: 10 25 67 94 98 
#> 3: (0.599555,0.4013588) [8]: 7 9 22 56 58 69 78 93 
#> 4: (0.1513727,0.214976) [5]: 39 88 90 92 96 
#> 5: (0,0.4158796) [6]: 27 31 59 73 76 99 
#> 6: (0.4637851,0.6676082) [5]: 35 36 51 77 81 
#> 7: (0.4954966,0.505603) [7]: 3 22 35 43 55 78 87 
#> 8: (0.2853152,0.6590554) [5]: 32 57 64 70 77 
#> 9: (0.7330552,0.5097642) [5]: 3 22 47 56 79 
#> 10: (0.7731085,0.8351012) [6]: 2 34 61 94 95 98 
#> 11: (0.8762532,0.7077129) [5]: 20 37 47 67 98 
#> 12: (0.16929,0.8739429) [7]: 1 24 40 45 53 62 83 
#> 13: (0.02712207,0.00701722) [4]: 31 39 73 90 
#> 14: (0.3162533,0.8880549) [6]: 21 33 40 50 60 83 
#> 15: (0.3990511,0.9966784) [4]: 60 66 74 100 
#> 16: (0.1902356,0.498108) [6]: 57 59 64 75 87 92 
#> 17: (0.4002736,0.3561961) [8]: 52 55 69 72 75 88 92 93 
#> 18: (0.05684929,0.7741667) [3]: 1 53 76 
#> 19: (0.385282,0.5828019) [4]: 36 64 77 87 
#> 20: (0.9782541,0.632544) [7]: 11 25 47 48 67 79 91 
#> 21: (0.2854416,0.8583275) [4]: 14 33 62 83 
#> 22: (0.6779848,0.5651355) [6]: 3 7 9 37 43 47 
#> 23: (0.7355183,0.2497103) [3]: 26 56 58 
#> 24: (0.1905255,0.9187572) [4]: 12 40 45 66 
#> 25: (0.983298,0.8670538) [6]: 2 20 49 67 91 94 
#> 26: (0.741785,0.2452302) [8]: 23 29 30 38 56 58 79 86 
#> 27: (0.0445066,0.400324) [4]: 5 59 92 99 
#> 28: (0.5282702,0.7688581) [7]: 50 51 63 65 81 85 97 
#> 29: (0.6956103,0.1155488) [7]: 26 30 41 44 46 82 86 
#> 30: (0.6882666,0.1911244) [5]: 26 29 44 58 69 
#> 31: (0.02407963,0.1608521) [5]: 5 13 73 90 99 
#> 32: (0.2204403,0.6619165) [4]: 8 53 57 70 
#> 33: (0.2964942,0.8562261) [4]: 14 21 50 62 
#> 34: (0.6356325,0.9265382) [6]: 10 42 61 63 65 95 
#> 35: (0.476548,0.5505481) [6]: 6 7 36 43 81 87 
#> 36: (0.4292056,0.5753563) [5]: 6 19 35 77 87 
#> 37: (0.706331,0.6862757) [7]: 11 22 43 47 89 95 98 
#> 38: (0.9510013,0.2413912) [4]: 26 48 79 86 
#> 39: (0.1747445,0.04031614) [7]: 4 13 46 73 84 90 96 
#> 40: (0.2116872,0.909765) [6]: 12 14 24 60 66 83 
#> 41: (0.6797859,0.06650706) [3]: 29 46 82 
#> 42: (0.496576,0.9972256) [6]: 34 61 63 74 97 100 
#> 43: (0.6409006,0.6103124) [6]: 7 22 35 37 81 89 
#> 44: (0.6596998,0.1688806) [5]: 29 30 68 69 82 
#> 45: (0.08954979,0.9093494) [5]: 1 12 24 66 76 
#> 46: (0.7661149,0.03311526) [7]: 29 39 41 73 82 84 86 
#> 47: (0.7702321,0.5919247) [6]: 9 11 20 22 37 79 
#> 48: (0.9935768,0.2336128) [5]: 20 38 79 86 91 
#> 49: (0.9731747,0.9061904) [4]: 25 61 91 94 
#> 50: (0.3857684,0.8183406) [9]: 14 28 33 51 60 62 71 80 85 
#> 51: (0.4585237,0.6987175) [6]: 6 28 50 77 80 81 
#> 52: (0.3110556,0.216553) [5]: 17 72 84 88 96 
#> 53: (0.1690225,0.7270162) [8]: 1 12 18 32 57 62 70 76 
#> 54: (0.5296455,0.2136229) [5]: 68 69 72 82 84 
#> 55: (0.491313,0.4539327) [6]: 7 17 75 78 87 93 
#> 56: (0.7799665,0.3299014) [6]: 3 9 23 26 58 79 
#> 57: (0.1988329,0.5666009) [7]: 8 16 32 53 59 64 76 
#> 58: (0.7133671,0.2489151) [6]: 3 23 26 30 56 69 
#> 59: (0.05842017,0.4617539) [6]: 5 16 27 57 76 92 
#> 60: (0.3504274,0.917609) [7]: 14 15 40 50 66 71 74 
#> 61: (0.8263363,0.9730613) [7]: 10 34 42 49 91 94 100 
#> 62: (0.2691997,0.8185511) [8]: 12 21 33 50 53 70 80 83 
#> 63: (0.5685185,0.9028005) [5]: 28 34 42 65 97 
#> 64: (0.3317467,0.5796776) [6]: 8 16 19 57 77 87 
#> 65: (0.59501,0.7722529) [6]: 28 34 63 81 89 95 
#> 66: (0.1860404,0.9954485) [6]: 15 24 40 45 60 100 
#> 67: (0.9501801,0.7099137) [5]: 2 11 20 25 98 
#> 68: (0.5406662,0.2114706) [4]: 44 54 69 82 
#> 69: (0.5428114,0.2886595) [9]: 3 17 30 44 54 58 68 72 93 
#> 70: (0.2740285,0.7207546) [6]: 8 32 53 62 77 80 
#> 71: (0.4438885,0.8663157) [4]: 50 60 74 85 
#> 72: (0.3679124,0.2350956) [5]: 17 52 54 69 84 
#> 73: (0.0208772,0) [5]: 5 13 31 39 46 
#> 74: (0.4633746,0.9435908) [7]: 15 42 60 71 85 97 100 
#> 75: (0.3866259,0.4357516) [5]: 16 17 55 87 92 
#> 76: (0.01279798,0.7497386) [7]: 1 5 18 45 53 57 59 
#> 77: (0.3734291,0.6665481) [8]: 6 8 19 36 51 64 70 80 
#> 78: (0.5582806,0.4054408) [4]: 3 7 55 93 
#> 79: (0.8585533,0.3484403) [7]: 9 20 26 38 47 48 56 
#> 80: (0.3813497,0.737166) [5]: 50 51 62 70 77 
#> 81: (0.5259508,0.6630007) [7]: 6 28 35 43 51 65 89 
#> 82: (0.5994304,0.08112133) [7]: 29 41 44 46 54 68 84 
#> 83: (0.2566229,0.8557811) [5]: 12 14 21 40 62 
#> 84: (0.2856011,0.07283984) [7]: 39 46 52 54 72 82 96 
#> 85: (0.4776096,0.8524778) [5]: 28 50 71 74 97 
#> 86: (0.922132,0.1023464) [5]: 26 29 38 46 48 
#> 87: (0.3974263,0.4826444) [8]: 7 16 19 35 36 55 64 75 
#> 88: (0.2079212,0.2439042) [5]: 4 17 52 92 96 
#> 89: (0.6713021,0.6853928) [5]: 37 43 65 81 95 
#> 90: (0.05174926,0.1599014) [6]: 4 13 31 39 92 99 
#> 91: (1,0.9529444) [5]: 20 25 48 49 61 
#> 92: (0.1431144,0.318903) [9]: 4 16 17 27 59 75 88 90 99 
#> 93: (0.5164927,0.3587757) [5]: 3 17 55 69 78 
#> 94: (0.8474753,0.8875262) [5]: 2 10 25 49 61 
#> 95: (0.7182904,0.8275265) [6]: 10 34 37 65 89 98 
#> 96: (0.2363562,0.0966282) [5]: 4 39 52 84 88 
#> 97: (0.5452768,0.9059435) [5]: 28 42 63 74 85 
#> 98: (0.8360389,0.7719734) [6]: 2 10 11 37 67 95 
#> 99: (0.02077116,0.3807185) [5]: 5 27 31 90 92 
#> 100: (0.4668071,1) [5]: 15 42 61 66 74 
#> number of nodes: 100 
#> number of arcs: 286 
#> number of boundary nodes: 11 
#> boundary nodes:  5 45 46 48 61 66 73 76 86 91 100 
#> number of triangles: 187 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.177105724 0.050371077 0.121679902 0.146926937 0.087247483 0.160592058 
#>                                                                         
#> 0.137356932 0.090987853 0.089136752 0.129921634 0.062640731 0.074892862 
#>                                                                         
#> 0.175125363 0.104853497 0.185879762 0.078094357 0.191953254 0.118212990 
#>                                                                         
#> 0.171547283 0.083468874 0.149995943 0.099119721 0.183382346 0.160883959 
#>                                                                         
#> 0.219891313 0.030785504 0.062985883 0.291576029 0.076302756 0.090972307 
#>                                                                         
#> 0.106896972 0.053448559 0.122826005 0.058448172 0.174881543 0.104264839 
#>                                                                         
#> 0.098178796 0.082680862 0.113220272 0.079584065 0.097734052 0.133555505 
#>                                                                         
#> 0.096353054 0.080743867 0.259008244 0.096058472 0.088108537 0.196296596 
#>                                                                         
#> 0.128979247 0.130195444 0.082305568 0.155461376 0.113953743 0.211950623 
#>                                                                         
#> 0.070238706 0.022993034 0.101412997 0.135698600 0.039296978 0.159061415 
#>                                                                         
#> 0.027685948 0.105371712 0.138961710 0.028930917 0.011250667 0.096980377 
#>                                                                         
#> 0.280803563 0.181884749 0.115897014 0.056513572 0.048120467 0.117171916 
#>                                                                         
#> 0.175238760 0.166070822 0.270095245 0.106699611 0.059804024 0.182917329 
#>                                                                         
#> 0.098432084 0.106254108 0.157868336 0.098007318 0.719360257 0.087499475 
#>                                                                         
#> 0.177247593 0.046523116 0.100257460 0.028655724 0.092881861 0.129319384 
#>                                                                         
#> 0.089439483 0.053903024 0.127077472 0.063006372 0.075931624 0.226906116 
#>                                                                         
#> 0.142010159 0.141443087 0.137073123 0.129248628 0.133208461 0.023453230 
#>                                                                         
#> 0.086240589 0.097980332 0.140010747 0.106461673 0.081673630 0.092561965 
#>                                                                         
#> 0.170687674 0.036205649 0.167361622 0.222979426 0.105192527 0.163161661 
#>                                                                         
#> 0.116568896 0.330644511 0.118708966 0.135104592 0.126449625 0.149699315 
#>                                                                         
#> 0.081002037 0.149449780 0.745973113 0.482153097 0.054689506 0.173459849 
#>                                                                         
#> 0.090140413 0.360536985 0.147066010 0.077218728 0.142983091 0.313938564 
#>                                                                         
#> 0.174824980 0.138568249 0.086221266 0.036449935 0.106611377 0.071060665 
#>                                                                         
#> 0.181939757 0.097915569 0.261052540 0.079694048 0.108568759 0.092218299 
#>                                                                         
#> 0.082601369 0.035411827 0.122888869 0.117471914 0.105402024 0.157141538 
#>                                                                         
#> 0.074982625 0.066088214 0.041524774 0.154485486 0.093341695 0.126178326 
#>                                                                         
#> 0.190242673 0.194047534 0.171971033 0.181587485 0.147292294 0.041475783 
#>                                                                         
#> 0.176216662 0.145699680 0.063518245 0.104254544 0.113833413 0.256161743 
#>                                                                         
#> 0.040837974 0.047146712 0.074279029 0.334104295 0.031551079 0.090362193 
#>                                                                         
#> 0.098519811 0.117753880 0.062336247 0.179182520 0.048776056 0.100721816 
#>                                                                         
#> 0.051839457 0.062723130 0.064937948 0.126597852 0.091960379 0.088431925 
#>                                                                         
#> 0.204389740 0.090180146 0.147872133 0.165107174 0.055338933 0.073959715 
#>                                                                         
#> 0.075806304 0.171269087 0.156994858 0.126706517 0.055504504 0.049591039 
#>                                                                         
#> 0.114237487 0.089201379 0.151331413 0.154855027 0.153864946 0.009393603 
#>                                                                         
#> 0.045180896 0.106796021 0.067803950 0.042814501 0.037463158 0.098450162 
#>                                                                         
#> 0.213014223 0.092823703 0.083401425 0.067837372 0.207766950 0.163337074 
#>                                                                         
#> 0.069030402 0.136736679 0.185296604 0.206052113 0.259849229 0.222730618 
#>                                                                         
#> 0.165710745 0.125349843 0.157728212 0.116247722 0.133568738 0.080717673 
#>                                                                         
#> 0.100891152 0.044550284 0.084580790 0.053626360 0.321137514 0.234564002 
#>                                                                         
#> 0.308290771 0.399225341 0.042964607 0.124412440 0.007703358 0.091685691 
#>                                                                         
#> 0.022165415 0.076822327 0.040424724 0.076102937 0.137656763 0.230088796 
#>                                                                         
#> 0.209251536 0.155843360 0.127878434 0.066826049 0.139858838 0.138136300 
#>                                                                         
#> 0.097768852 0.150848525 0.098915489 0.105882847 0.102155891 0.051531559 
#>                                                                         
#> 0.108472042 0.064294983 0.175129425 0.082956374 0.155991738 0.071188204 
#>                                                                         
#> 0.159543535 0.141756257 0.035040004 0.043280256 0.591414273 0.115529095 
#>                                                                         
#> 0.063079518 0.103461037 0.029897839 0.126423530 0.161348128 0.081294796 
#>                                                                         
#> 0.075362838 0.145949985 0.141286767 0.076182913 0.163152301 0.281740233 
#>                                                 
#> 0.149755128 0.011228915 0.115607754 0.416403242 
#> 
#> $alpha
#> [1] 0.2781132
#> 
#> attr(,"class")
#> [1] "scree"