Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Voronoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2          mx1
#>    59   48 0.622109019 0.953009678 0.606601943 0.960550784  0.594038261
#>    51    1 0.591678714 0.791011495 0.640441321 0.906091940  0.577621604
#>     1   48 0.640441321 0.906091940 0.606601943 0.960550784  0.555234583
#>    90   30 0.585739110 1.000000000 0.454544086 0.875618479  0.518715742
#>    30   47 0.454544086 0.875618479 0.527389794 0.808938454  0.541859871
#>    90   59 0.585739110 1.000000000 0.622109019 0.953009678  0.701122182
#>    59    1 0.622109019 0.953009678 0.640441321 0.906091940  0.683637589
#>     1   65 0.640441321 0.906091940 0.687757834 0.721254497  0.691531646
#>    47   80 0.527389794 0.808938454 0.416984473 0.798301035  0.479251070
#>    60   80 0.389009658 0.827916501 0.416984473 0.798301035  0.430799905
#>    51   65 0.591678714 0.791011495 0.687757834 0.721254497  0.685465667
#>    68   65 0.645635090 0.687824029 0.687757834 0.721254497  0.633044558
#>    14    1 0.742636369 0.932293199 0.640441321 0.906091940  0.709283763
#>    34   60 0.401334126 0.924426530 0.389009658 0.827916501  0.404919442
#>    32   38 0.369038491 0.952830732 0.331397344 0.885042588  0.362075991
#>    34   90 0.401334126 0.924426530 0.585739110 1.000000000  0.465689896
#>    35   22 0.395971729 0.690738900 0.553216947 0.573473036  0.516740705
#>    11   65 0.790875433 0.814901017 0.687757834 0.721254497  0.718233056
#>    94   66 0.704947950 0.627519212 0.727947963 0.715531901  0.701727257
#>    38   95 0.331397344 0.885042588 0.327312973 0.843438986  0.364480015
#>    38   60 0.331397344 0.885042588 0.389009658 0.827916501  0.381464908
#>    60    4 0.389009658 0.827916501 0.310649092 0.739408802  0.361182753
#>    68   51 0.645635090 0.687824029 0.591678714 0.791011495  0.529673197
#>     4   35 0.310649092 0.739408802 0.395971729 0.690738900  0.373753743
#>    68   94 0.645635090 0.687824029 0.704947950 0.627519212  0.691354275
#>    22   94 0.553216947 0.573473036 0.704947950 0.627519212  0.625989183
#>     3    4 0.292249979 0.847325267 0.310649092 0.739408802  0.304065712
#>    31    3 0.218975721 0.887475286 0.292249979 0.847325267  0.275566447
#>    46   11 0.922934924 0.974567020 0.790875433 0.814901017  0.872966556
#>    37   11 0.853595941 0.742260701 0.790875433 0.814901017  0.786725156
#>    60   95 0.389009658 0.827916501 0.327312973 0.843438986  0.345940827
#>    90   14 0.585739110 1.000000000 0.742636369 0.932293199  0.771498885
#>    46   14 0.922934924 0.974567020 0.742636369 0.932293199  0.844056090
#>    22   68 0.553216947 0.573473036 0.645635090 0.687824029  0.526386059
#>    82   22 0.441099207 0.572081348 0.553216947 0.573473036  0.496064286
#>    44   96 0.879025274 0.671200181 0.960238047 0.821856360  0.942729331
#>    37   66 0.853595941 0.742260701 0.727947963 0.715531901  0.800497318
#>   100   44 0.928961922 0.622069409 0.879025274 0.671200181  0.837641944
#>    32   31 0.369038491 0.952830732 0.218975721 0.887475286  0.276776099
#>    64   82 0.379375279 0.601311898 0.441099207 0.572081348  0.434240812
#>    31   20 0.218975721 0.887475286 0.207015697 0.891825850  0.169111331
#>    90   46 0.585739110 1.000000000 0.922934924 0.974567020  1.383919640
#>    46   96 0.922934924 0.974567020 0.960238047 0.821856360  0.943916389
#>     4   31 0.310649092 0.739408802 0.218975721 0.887475286  0.198493580
#>    27   19 0.362577812 0.587190021 0.379253737 0.537947083  0.391696142
#>    71   27 0.261555328 0.637436881 0.362577812 0.587190021  0.328319388
#>     4   77 0.310649092 0.739408802 0.244520269 0.664921825  0.301312607
#>    64   35 0.379375279 0.601311898 0.395971729 0.690738900  0.331815316
#>    19   82 0.379253737 0.537947083 0.441099207 0.572081348  0.402134046
#>    52   19 0.277815730 0.560262400 0.379253737 0.537947083  0.328323572
#>    58  100 0.963430047 0.568269173 0.928961922 0.622069409  0.904065959
#>    71   52 0.261555328 0.637436881 0.277815730 0.560262400  0.309547289
#>    84   52 0.225292721 0.547798783 0.277815730 0.560262400  0.242288537
#>    58   33 0.963430047 0.568269173 1.000000000 0.313048880  2.930094455
#>    96  100 0.960238047 0.821856360 0.928961922 0.622069409  1.104928046
#>    44   94 0.879025274 0.671200181 0.704947950 0.627519212  0.810885746
#>    77   84 0.244520269 0.664921825 0.225292721 0.547798783  0.192071044
#>    56   84 0.102377463 0.656266127 0.225292721 0.547798783  0.176164159
#>    22   72 0.553216947 0.573473036 0.500289322 0.391768101  0.617135106
#>    19   99 0.379253737 0.537947083 0.269695689 0.476748912  0.351084425
#>    23   94 0.910230911 0.509134729 0.704947950 0.627519212  0.782156434
#>    26   84 0.244144071 0.467437723 0.225292721 0.547798783  0.185121748
#>    20   56 0.207015697 0.891825850 0.102377463 0.656266127  0.166869554
#>    72   19 0.500289322 0.391768101 0.379253737 0.537947083  0.442213544
#>    42   99 0.391140581 0.386841572 0.269695689 0.476748912  0.332599297
#>    26   99 0.244144071 0.467437723 0.269695689 0.476748912  0.243245422
#>    74   88 0.106492324 0.510926607 0.132720865 0.535387468  0.109212403
#>    72   81 0.500289322 0.391768101 0.406841965 0.363130727  0.459922396
#>    72   42 0.500289322 0.391768101 0.391140581 0.386841572  0.444986482
#>    99   21 0.269695689 0.476748912 0.226396680 0.425580679  0.278739492
#>     9   70 0.859662118 0.387574574 0.951119352 0.335153149  0.941715899
#>    23   61 0.910230911 0.509134729 0.739330439 0.402583430  0.810028571
#>    75    9 0.819600039 0.368483373 0.859662118 0.387574574  0.806136476
#>    86   79 0.422131876 0.318882213 0.436872847 0.252075207  0.509819398
#>    88   21 0.132720865 0.535387468 0.226396680 0.425580679  0.171451990
#>    67    6 0.079929295 0.664357598 0.018260031 0.687602408  0.035425597
#>    67   20 0.079929295 0.664357598 0.207015697 0.891825850  0.097305570
#>    43   61 0.643086998 0.186352907 0.739330439 0.402583430  0.623084956
#>    86   81 0.422131876 0.318882213 0.406841965 0.363130727  0.343442815
#>    81   42 0.406841965 0.363130727 0.391140581 0.386841572  0.344810531
#>    21   16 0.226396680 0.425580679 0.043804776 0.380150328  0.134393572
#>    88   56 0.132720865 0.535387468 0.102377463 0.656266127  0.105558559
#>    67   10 0.079929295 0.664357598 0.003736775 0.599920661  0.069294234
#>    73   10 0.085751568 0.532454546 0.003736775 0.599920661 -0.013159486
#>    79   43 0.436872847 0.252075207 0.643086998 0.186352907  0.559551713
#>    43   25 0.643086998 0.186352907 0.812849365 0.239856564  0.704138880
#>    81   98 0.406841965 0.363130727 0.287972482 0.373668329  0.342865596
#>    98   29 0.287972482 0.373668329 0.226964112 0.296649156  0.270290336
#>    70   75 0.951119352 0.335153149 0.819600039 0.368483373  0.872535493
#>    98   21 0.287972482 0.373668329 0.226396680 0.425580679  0.224711223
#>    74   73 0.106492324 0.510926607 0.085751568 0.532454546  0.037401392
#>    86   91 0.422131876 0.318882213 0.284793899 0.263939325  0.358158302
#>    91   98 0.284793899 0.263939325 0.287972482 0.373668329  0.297752734
#>    33    7 1.000000000 0.313048880 0.972668565 0.971972912  3.020032389
#>    70   33 0.951119352 0.335153149 1.000000000 0.313048880  1.030505134
#>    45   33 0.936090170 0.226835318 1.000000000 0.313048880  0.955330667
#>    25   75 0.812849365 0.239856564 0.819600039 0.368483373  0.746456036
#>    24   17 0.335310628 0.206672818 0.388830962 0.211196877  0.356386647
#>    24   53 0.335310628 0.206672818 0.336675803 0.195459590  0.362463517
#>    53   91 0.336675803 0.195459590 0.284793899 0.263939325  0.256113073
#>    17    2 0.388830962 0.211196877 0.452004759 0.148213379  0.440176313
#>    29   54 0.226964112 0.296649156 0.073090459 0.169951187  0.186222120
#>    29   21 0.226964112 0.296649156 0.226396680 0.425580679  0.169302660
#>    74   16 0.106492324 0.510926607 0.043804776 0.380150328  0.131164102
#>    25   45 0.812849365 0.239856564 0.936090170 0.226835318  0.880430277
#>    25   89 0.812849365 0.239856564 0.806940503 0.129241906  0.846300094
#>    69   54 0.092554032 0.321393023 0.073090459 0.169951187  0.146588311
#>    79    2 0.436872847 0.252075207 0.452004759 0.148213379  0.530593290
#>    53   50 0.336675803 0.195459590 0.359208741 0.164233162  0.365923625
#>    50    2 0.359208741 0.164233162 0.452004759 0.148213379  0.407407747
#>    13   50 0.333939191 0.148769619 0.359208741 0.164233162  0.337083249
#>    97   43 0.772375036 0.129587811 0.643086998 0.186352907  0.690263331
#>    40   93 0.308777572 0.126934063 0.445426809 0.080349835  0.369753396
#>    69   16 0.092554032 0.321393023 0.043804776 0.380150328 -0.001299798
#>    25   97 0.812849365 0.239856564 0.772375036 0.129587811  0.790220030
#>    93    2 0.445426809 0.080349835 0.452004759 0.148213379  0.399194825
#>    91   55 0.284793899 0.263939325 0.264183085 0.273584690  0.235122407
#>    41   63 0.901881312 0.117206325 0.974618732 0.224656090  0.951737643
#>    85   12 0.689719117 0.035288251 0.581236117 0.124551505  0.589650739
#>    43   12 0.643086998 0.186352907 0.581236117 0.124551505  0.659031259
#>    13   93 0.333939191 0.148769619 0.445426809 0.080349835  0.380948965
#>    40   13 0.308777572 0.126934063 0.333939191 0.148769619  0.289283763
#>    83   40 0.191914383 0.068506735 0.308777572 0.126934063  0.210614309
#>    89   97 0.806940503 0.129241906 0.772375036 0.129587811  0.788917376
#>    12    2 0.581236117 0.124551505 0.452004759 0.148213379  0.511460288
#>    87   41 0.887541456 0.130682502 0.901881312 0.117206325  0.935151584
#>    39   41 0.880191344 0.044037843 0.901881312 0.117206325  0.862068352
#>    92   49 0.688508422 0.005167514 0.854197313 0.016548455  0.770935165
#>    54   83 0.073090459 0.169951187 0.191914383 0.068506735  0.188606639
#>    28   54 0.000000000 0.391136120 0.073090459 0.169951187 -0.026473840
#>    93   78 0.445426809 0.080349835 0.293882672 0.055707170  0.369824602
#>    97   85 0.772375036 0.129587811 0.689719117 0.035288251  0.774098382
#>    12   93 0.581236117 0.124551505 0.445426809 0.080349835  0.540789113
#>    92   12 0.688508422 0.005167514 0.581236117 0.124551505  0.557097813
#>    89   39 0.806940503 0.129241906 0.880191344 0.044037843  0.847947954
#>    15   18 0.249398283 0.047686234 0.222851881 0.034042737  0.239172511
#>    78   40 0.293882672 0.055707170 0.308777572 0.126934063  0.263052636
#>    15   40 0.249398283 0.047686234 0.308777572 0.126934063  0.241451545
#>    63   39 0.974618732 0.224656090 0.880191344 0.044037843  1.332061524
#>    39   49 0.880191344 0.044037843 0.854197313 0.016548455  0.876104971
#>    57    5 0.236228625 0.018857338 0.426926355 0.015664019  0.330205865
#>    15   83 0.249398283 0.047686234 0.191914383 0.068506735  0.223389878
#>    57   62 0.236228625 0.018857338 0.140801819 0.001443768  0.319443711
#>    78   15 0.293882672 0.055707170 0.249398283 0.047686234  0.277848455
#>    57   18 0.236228625 0.018857338 0.222851881 0.034042737  0.191634152
#>    36   62 0.097225613 0.072531109 0.140801819 0.001443768  0.143790379
#>    76   36 0.056949761 0.059175380 0.097225613 0.072531109  0.060727734
#>    76   62 0.056949761 0.059175380 0.140801819 0.001443768  0.092125437
#>    28   76 0.000000000 0.391136120 0.056949761 0.059175380 -0.303325681
#>     8   92 0.870622338 0.000000000 0.688508422 0.005167514  0.485192784
#>    62   92 0.140801819 0.001443768 0.688508422 0.005167514  0.435341210
#>    62   18 0.140801819 0.001443768 0.222851881 0.034042737  0.178725421
#>    47    1 0.527389794 0.808938454 0.640441321 0.906091940  0.577621604
#>    47   48 0.527389794 0.808938454 0.606601943 0.960550784  0.555234583
#>    48   30 0.606601943 0.960550784 0.454544086 0.875618479  0.518715742
#>    14   59 0.742636369 0.932293199 0.622109019 0.953009678  0.701122182
#>    35   80 0.395971729 0.690738900 0.416984473 0.798301035  0.479251070
#>    80   30 0.416984473 0.798301035 0.454544086 0.875618479  0.470605603
#>    11    1 0.790875433 0.814901017 0.640441321 0.906091940  0.709283763
#>    30   60 0.454544086 0.875618479 0.389009658 0.827916501  0.404919442
#>    34   38 0.401334126 0.924426530 0.331397344 0.885042588  0.362075991
#>    32   90 0.369038491 0.952830732 0.585739110 1.000000000  0.465689896
#>    47   22 0.527389794 0.808938454 0.553216947 0.573473036  0.516740705
#>    66   65 0.727947963 0.715531901 0.687757834 0.721254497  0.718233056
#>    80    4 0.416984473 0.798301035 0.310649092 0.739408802  0.361182753
#>    47   51 0.527389794 0.808938454 0.591678714 0.791011495  0.529673197
#>    65   94 0.687757834 0.721254497 0.704947950 0.627519212  0.691354275
#>    95    4 0.327312973 0.843438986 0.310649092 0.739408802  0.304065712
#>     3   38 0.292249979 0.847325267 0.331397344 0.885042588  0.312059910
#>    96   11 0.960238047 0.821856360 0.790875433 0.814901017  0.872966556
#>    66   11 0.727947963 0.715531901 0.790875433 0.814901017  0.786725156
#>    11   14 0.790875433 0.814901017 0.742636369 0.932293199  0.844056090
#>    47   68 0.527389794 0.808938454 0.645635090 0.687824029  0.526386059
#>    37   96 0.853595941 0.742260701 0.960238047 0.821856360  0.942729331
#>    44   66 0.879025274 0.671200181 0.727947963 0.715531901  0.800497318
#>    23   44 0.910230911 0.509134729 0.879025274 0.671200181  0.837641944
#>    71    4 0.261555328 0.637436881 0.310649092 0.739408802  0.326903099
#>    38   31 0.331397344 0.885042588 0.218975721 0.887475286  0.276776099
#>    35   82 0.395971729 0.690738900 0.441099207 0.572081348  0.434240812
#>    77   20 0.244520269 0.664921825 0.207015697 0.891825850  0.169111331
#>     7   46 0.972668565 0.971972912 0.922934924 0.974567020  1.383919640
#>     7   96 0.972668565 0.971972912 0.960238047 0.821856360  0.943916389
#>    77   31 0.244520269 0.664921825 0.218975721 0.887475286  0.198493580
#>    20   32 0.207015697 0.891825850 0.369038491 0.952830732  0.255910538
#>    64   19 0.379375279 0.601311898 0.379253737 0.537947083  0.391696142
#>    64   27 0.379375279 0.601311898 0.362577812 0.587190021  0.328319388
#>    71   77 0.261555328 0.637436881 0.244520269 0.664921825  0.301312607
#>    71   35 0.261555328 0.637436881 0.395971729 0.690738900  0.331815316
#>    72   82 0.500289322 0.391768101 0.441099207 0.572081348  0.498173835
#>    61   22 0.739330439 0.402583430 0.553216947 0.573473036  0.662751126
#>    23  100 0.910230911 0.509134729 0.928961922 0.622069409  0.904065959
#>    27   52 0.362577812 0.587190021 0.277815730 0.560262400  0.309547289
#>    96   33 0.960238047 0.821856360 1.000000000 0.313048880  2.930094455
#>    99   84 0.269695689 0.476748912 0.225292721 0.547798783  0.259658982
#>    71   84 0.261555328 0.637436881 0.225292721 0.547798783  0.192071044
#>    61   72 0.739330439 0.402583430 0.500289322 0.391768101  0.617135106
#>    61   94 0.739330439 0.402583430 0.704947950 0.627519212  0.782156434
#>    99   52 0.269695689 0.476748912 0.277815730 0.560262400  0.320797051
#>    88   84 0.132720865 0.535387468 0.225292721 0.547798783  0.185121748
#>    77   56 0.244520269 0.664921825 0.102377463 0.656266127  0.166869554
#>    42   19 0.391140581 0.386841572 0.379253737 0.537947083  0.442213544
#>    98   99 0.287972482 0.373668329 0.269695689 0.476748912  0.332599297
#>    73   88 0.085751568 0.532454546 0.132720865 0.535387468  0.109212403
#>    23   70 0.910230911 0.509134729 0.951119352 0.335153149  0.941715899
#>     9   61 0.859662118 0.387574574 0.739330439 0.402583430  0.810028571
#>    72   79 0.500289322 0.391768101 0.436872847 0.252075207  0.509819398
#>    26   21 0.244144071 0.467437723 0.226396680 0.425580679  0.171451990
#>    10    6 0.003736775 0.599920661 0.018260031 0.687602408  0.035425597
#>     6   20 0.018260031 0.687602408 0.207015697 0.891825850  0.097305570
#>    91   81 0.284793899 0.263939325 0.406841965 0.363130727  0.343442815
#>    98   42 0.287972482 0.373668329 0.391140581 0.386841572  0.344810531
#>    73   56 0.085751568 0.532454546 0.102377463 0.656266127  0.105558559
#>    56   10 0.102377463 0.656266127 0.003736775 0.599920661  0.069294234
#>    28   10 0.000000000 0.391136120 0.003736775 0.599920661 -0.013159486
#>    23   58 0.910230911 0.509134729 0.963430047 0.568269173  1.037005493
#>    72   43 0.500289322 0.391768101 0.643086998 0.186352907  0.559551713
#>    61   25 0.739330439 0.402583430 0.812849365 0.239856564  0.704138880
#>    55   29 0.264183085 0.273584690 0.226964112 0.296649156  0.270290336
#>    28   73 0.000000000 0.391136120 0.085751568 0.532454546  0.037401392
#>    17   91 0.388830962 0.211196877 0.284793899 0.263939325  0.358158302
#>    55   98 0.264183085 0.273584690 0.287972482 0.373668329  0.297752734
#>    23   33 0.910230911 0.509134729 1.000000000 0.313048880  1.030505134
#>    61   75 0.739330439 0.402583430 0.819600039 0.368483373  0.746456036
#>    12   79 0.581236117 0.124551505 0.436872847 0.252075207  0.542028407
#>    17   86 0.388830962 0.211196877 0.422131876 0.318882213  0.376911602
#>    17   53 0.388830962 0.211196877 0.336675803 0.195459590  0.362463517
#>    40   91 0.308777572 0.126934063 0.284793899 0.263939325  0.256113073
#>    91   24 0.284793899 0.263939325 0.335310628 0.206672818  0.260869292
#>    55   54 0.264183085 0.273584690 0.073090459 0.169951187  0.186222120
#>    69   21 0.092554032 0.321393023 0.226396680 0.425580679  0.169302660
#>    70   45 0.951119352 0.335153149 0.936090170 0.226835318  0.880430277
#>    87   89 0.887541456 0.130682502 0.806940503 0.129241906  0.846300094
#>    17   50 0.388830962 0.211196877 0.359208741 0.164233162  0.365923625
#>    63   45 0.974618732 0.224656090 0.936090170 0.226835318  0.958264354
#>    85   43 0.689719117 0.035288251 0.643086998 0.186352907  0.690263331
#>    28   16 0.000000000 0.391136120 0.043804776 0.380150328 -0.001299798
#>    28   74 0.000000000 0.391136120 0.106492324 0.510926607  0.041025626
#>    40   55 0.308777572 0.126934063 0.264183085 0.273584690  0.235122407
#>    87   25 0.887541456 0.130682502 0.812849365 0.239856564  0.870885782
#>    50   93 0.359208741 0.164233162 0.445426809 0.080349835  0.380948965
#>    53   13 0.336675803 0.195459590 0.333939191 0.148769619  0.289283763
#>    49   97 0.854197313 0.016548455 0.772375036 0.129587811  0.788917376
#>    45   41 0.936090170 0.226835318 0.901881312 0.117206325  0.935151584
#>    85   49 0.689719117 0.035288251 0.854197313 0.016548455  0.770935165
#>    55   83 0.264183085 0.273584690 0.191914383 0.068506735  0.188606639
#>     5   78 0.426926355 0.015664019 0.293882672 0.055707170  0.369824602
#>     5   93 0.426926355 0.015664019 0.445426809 0.080349835  0.540789113
#>     5   12 0.426926355 0.015664019 0.581236117 0.124551505  0.557097813
#>    87   39 0.887541456 0.130682502 0.880191344 0.044037843  0.847947954
#>    49   89 0.854197313 0.016548455 0.806940503 0.129241906  0.824730604
#>     8   39 0.870622338 0.000000000 0.880191344 0.044037843  1.332061524
#>     8   49 0.870622338 0.000000000 0.854197313 0.016548455  0.876104971
#>    18   83 0.222851881 0.034042737 0.191914383 0.068506735  0.223389878
#>    36   54 0.097225613 0.072531109 0.073090459 0.169951187  0.147380913
#>     5   62 0.426926355 0.015664019 0.140801819 0.001443768  0.319443711
#>    57   15 0.236228625 0.018857338 0.249398283 0.047686234  0.277848455
#>    83   62 0.191914383 0.068506735 0.140801819 0.001443768  0.143790379
#>    54   76 0.073090459 0.169951187 0.056949761 0.059175380 -0.303325681
#>     5   92 0.426926355 0.015664019 0.688508422 0.005167514  0.435341210
#>    90   48 0.585739110 1.000000000 0.606601943 0.960550784  0.636033134
#>    34   30 0.401334126 0.924426530 0.454544086 0.875618479  0.490895776
#>    35   47 0.395971729 0.690738900 0.527389794 0.808938454  0.479251070
#>    34   32 0.401334126 0.924426530 0.369038491 0.952830732  0.362075991
#>    95    3 0.327312973 0.843438986 0.292249979 0.847325267  0.304065712
#>    37   44 0.853595941 0.742260701 0.879025274 0.671200181  0.942729331
#>     7   90 0.972668565 0.971972912 0.585739110 1.000000000  1.383919640
#>    64   71 0.379375279 0.601311898 0.261555328 0.637436881  0.328319388
#>    96   58 0.960238047 0.821856360 0.963430047 0.568269173  2.930094455
#>    88   26 0.132720865 0.535387468 0.244144071 0.467437723  0.185121748
#>    67   56 0.079929295 0.664357598 0.102377463 0.656266127  0.135260951
#>    86   72 0.422131876 0.318882213 0.500289322 0.391768101  0.459922396
#>    23    9 0.910230911 0.509134729 0.859662118 0.387574574  0.941715899
#>    74   21 0.106492324 0.510926607 0.226396680 0.425580679  0.168334031
#>    25   70 0.812849365 0.239856564 0.951119352 0.335153149  0.872535493
#>    17   79 0.388830962 0.211196877 0.436872847 0.252075207  0.376911602
#>    40   53 0.308777572 0.126934063 0.336675803 0.195459590  0.256113073
#>    69   29 0.092554032 0.321393023 0.226964112 0.296649156  0.169302660
#>    63   33 0.974618732 0.224656090 1.000000000 0.313048880  0.958264354
#>    28   69 0.000000000 0.391136120 0.092554032 0.321393023 -0.001299798
#>    92   85 0.688508422 0.005167514 0.689719117 0.035288251  0.589650739
#>    87   45 0.887541456 0.130682502 0.936090170 0.226835318  0.870885782
#>     8   63 0.870622338 0.000000000 0.974618732 0.224656090  1.332061524
#>    78   57 0.293882672 0.055707170 0.236228625 0.018857338  0.330205865
#>    36   83 0.097225613 0.072531109 0.191914383 0.068506735  0.147380913
#>    62    8 0.140801819 0.001443768 0.870622338 0.000000000  0.485192784
#>            my1          mx2           my2 bp1 bp2
#>    0.915001125   0.63603313   1.001356849   0   0
#>    0.864839083   0.68546567   0.819142721   0   0
#>    0.890889480   0.59403826   0.915001125   0   0
#>    0.939313202   0.49089578   0.968657120   0   0
#>    0.897877300   0.47060560   0.820034383   0   0
#>    1.051734941   0.63603313   1.001356849   0   0
#>    0.950010526   0.59403826   0.915001125   0   0
#>    0.820695550   0.68546567   0.819142721   0   0
#>    0.730303443   0.47060560   0.820034383   0   0
#>    0.839371373   0.36118275   0.773610903   0   0
#>    0.819142721   0.63304456   0.746941018   0   0
#>    0.746941018   0.69135428   0.673470158   0   0
#>    0.849980525   0.68363759   0.950010526   0   0
#>    0.874926739   0.38146491   0.877921917   0   0
#>    0.912352153   0.27677610   0.959717158   0   0
#>    1.030161373   0.49089578   0.968657120   0   0
#>    0.688621261   0.49606429   0.660895647   0   0
#>    0.791293653   0.69153165   0.820695550   0   0
#>    0.675372459   0.79133524   0.651955555   0   0
#>    0.860792457   0.31205991   0.865938722   0   0
#>    0.877921917   0.36448001   0.860792457   0   0
#>    0.773610903   0.34594083   0.787105390   0   0
#>    0.692888484   0.63304456   0.746941018   0   0
#>    0.750912817   0.32690310   0.668779500   0   0
#>    0.673470158   0.62598918   0.609180259   0   0
#>    0.609180259   0.66275113   0.505973576   0   0
#>    0.793813077   0.20614338   0.777117903   0   0
#>    0.903815826   0.20614338   0.777117903   0   0
#>    0.881449679   0.84405609   0.905361479   0   0
#>    0.747919669   0.87530919   0.824406608   0   0
#>    0.787105390   0.36448001   0.860792457   0   0
#>    1.214819175   0.70112218   1.051734941   0   0
#>    0.905361479   0.77149889   1.214819175   0   0
#>    0.689679217   0.62598918   0.609180259   0   0
#>    0.660895647   0.49817383   0.490945243   0   0
#>    0.734077232   0.97363417   0.717417659   0   0
#>    0.683178854   0.78672516   0.747919669   0   0
#>    0.579194795   0.97363417   0.717417659   0   0
#>    0.959717158   0.25591054   1.007626615   0   0
#>    0.637383133   0.40213405   0.569585719   0   0
#>    0.769009158   0.25591054   1.007626615   0   0
#>    9.334422609   0.77149889   1.214819175   0   0
#>    0.898780823   0.87296656   0.881449679   0   0
#>    0.772381635   0.20614338   0.777117903   0   0
#>    0.569605740   0.32832357   0.548144872   0   0
#>    0.644990122   0.30954729   0.607248379   0   0
#>    0.681099888   0.19849358   0.772381635   0   0
#>    0.656391928   0.43424081   0.637383133   0   0
#>    0.569585719   0.45361162   0.476317130   0   0
#>    0.548144872   0.32079705   0.513931813   0   0
#>    0.568177905   1.10492805   0.696863945   0   0
#>    0.607248379   0.24228854   0.593077187   0   0
#>    0.593077187   0.25965898   0.519876285   0   0
#>    0.719838019   1.03700549   0.448581476   0   0
#>    0.696863945   0.97363417   0.717417659   0   0
#>    0.574042900   0.79133524   0.651955555   0   0
#>    0.613392402   0.17616416   0.616003761   0   0
#>    0.616003761   0.16997128   0.608985989   0   0
#>    0.456293804   0.49817383   0.490945243   0   0
#>    0.459710820   0.32079705   0.513931813   0   0
#>    0.524225237   0.81088575   0.574042900   0   0
#>    0.495983715   0.24324542   0.509618550   0   0
#>    0.768638619   0.13526095   0.782679510   0   0
#>    0.466879570   0.45361162   0.476317130   0   0
#>    0.434741509   0.35108442   0.459710820   0   0
#>    0.509618550   0.27190296   0.430976981   0   0
#>    0.534302372   0.16833403   0.470908283   0   0
#>    0.356706531   0.44498648   0.405444298   0   0
#>    0.405444298   0.44221354   0.466879570   0   0
#>    0.425191849   0.27190296   0.430976981   0   0
#>    0.424738695   0.87382029   0.306284362   0   0
#>    0.479520373   0.78215643   0.524225237   0   0
#>    0.448315981   0.87382029   0.306284362   0   0
#>    0.303200670   0.37691160   0.273874555   0   0
#>    0.473568204   0.16833403   0.470908283   0   0
#>    0.639715511   0.09730557   0.803885164   0   0
#>    0.803885164   0.13526095   0.782679510   0   0
#>    0.324789814   0.70413888   0.288712999   0   0
#>    0.316457441   0.45992240   0.356706531   0   0
#>    0.339107520   0.44498648   0.405444298   0   0
#>    0.405707677   0.13116410   0.418687439   0   0
#>    0.592816872   0.16997128   0.608985989   0   0
#>    0.599668033   0.03542560   0.639715511   0   0
#>    0.495797356   0.07052011   0.597521970   0   0
#>    0.280623658   0.54202841   0.225641504   0   0
#>    0.288712999   0.72962716   0.207840952   0   0
#>    0.317167669   0.34481053   0.339107520   0   0
#>    0.325002158   0.22471122   0.361106251   0   0
#>    0.301214630   0.87382029   0.306284362   0   0
#>    0.361106251   0.27873949   0.425191849   0   0
#>    0.465117177   0.10921240   0.534302372   0   0
#>    0.279673878   0.34344282   0.316457441   0   0
#>    0.318474479   0.34286560   0.317167669   0   0
#>    0.726866429   4.53505143   0.789707733   0   1
#>    0.445605579   0.95533067   0.279367255   0   0
#>    0.279367255   0.95826435   0.277192523   0   0
#>    0.307831612   0.87253549   0.301214630   0   0
#>    0.276179203   0.36246352   0.204288880   0   0
#>    0.204288880   0.26086929   0.191920103   0   0
#>    0.188316676   0.26086929   0.191920103   0   0
#>    0.199523279   0.40740775   0.166655706   0   0
#>    0.189341838   0.14658831   0.237476777   0   0
#>    0.360862396   0.22471122   0.361106251   0   0
#>    0.418687439   0.04102563   0.461895275   0   0
#>    0.289759742   0.87088578   0.199424975   0   0
#>    0.182604528   0.79022003   0.185600238   0   0
#>    0.237476777  -0.02647384   0.259719031   0   0
#>    0.212696374   0.44017631   0.199523279   0   0
#>    0.192821690   0.33708325   0.172010524   0   0
#>    0.166655706   0.39919482   0.119081627   0   0
#>    0.172010524   0.38094896   0.100327926   0   0
#>    0.118186002   0.72962716   0.207840952   0   0
#>    0.082085142   0.36813903   0.077349601   0   0
#>    0.293126769   0.13439357   0.405707677   0   0
#>    0.185600238   0.72962716   0.207840952   0   0
#>    0.119081627   0.51146029   0.108199841   0   0
#>    0.184642122   0.29775273   0.318474479   0   0
#>    0.161800850   1.21237291  -0.014634503   0   0
#>    0.024225771   0.65903126   0.108544980   0   0
#>    0.108544980   0.54202841   0.225641504   0   0
#>    0.100327926   0.36975340   0.082085142   0   0
#>    0.174812167   0.36975340   0.082085142   0   0
#>    0.177189539   0.24145154   0.115510563   0   0
#>    0.055429050   0.79022003   0.185600238   0   0
#>    0.108199841   0.53059329   0.212696374   0   0
#>    0.166976398   0.86206835   0.089209312   0   0
#>    0.089209312   1.21237291  -0.014634503   0   0
#>    0.016939094   0.77733072  -0.076170294   0   0
#>    0.184944959   0.14738091   0.136656438   0   0
#>    0.259719031  -0.30332568   0.168233469   0   0
#>    0.066983911   0.36813903   0.077349601   0   0
#>    0.044702486   0.69026333   0.118186002   0   0
#>    0.018087241   0.51146029   0.108199841   0   0
#>   -0.005024612   0.58965074   0.024225771   0   0
#>    0.090407153   0.82473060   0.070446949   0   0
#>    0.034935047   0.22338988   0.065643605   0   0
#>    0.099325168   0.36813903   0.077349601   0   0
#>    0.115510563   0.26305264   0.099325168   0   0
#>   -0.077207821   1.21237291  -0.014634503   0   0
#>    0.021867225   0.82473060   0.070446949   0   0
#>   -0.064649650   0.31944371  -0.707340936   0   0
#>    0.065643605   0.24145154   0.115510563   0   0
#>   -0.707340936   0.19163415  -0.006941262   0   0
#>    0.017267047   0.26305264   0.099325168   0   0
#>   -0.006941262   0.23917251   0.034935047   0   0
#>    0.052175417   0.09212544   0.020505050   0   0
#>    0.115188708   0.09212544   0.020505050   0   0
#>    0.020505050  -5.47821594  -8.070115892   0   1
#>    0.168233469  -3.22960569  -0.333786579   0   1
#>  -10.371718208   0.77733072  -0.076170294   0   0
#>   -3.039305394   0.48519278 -10.371718208   0   0
#>    0.025549402   0.19163415  -0.006941262   0   0
#>    0.864839083   0.55523458   0.890889480   0   0
#>    0.890889480   0.54185987   0.897877300   0   0
#>    0.939313202   0.54185987   0.897877300   0   0
#>    1.051734941   0.68363759   0.950010526   0   0
#>    0.730303443   0.37375374   0.750912817   0   0
#>    0.820034383   0.43079990   0.839371373   0   0
#>    0.849980525   0.69153165   0.820695550   0   0
#>    0.874926739   0.43079990   0.839371373   0   0
#>    0.912352153   0.38146491   0.877921917   0   0
#>    1.030161373  -0.49334503   5.436069529   0   1
#>    0.688621261   0.52638606   0.689679217   0   0
#>    0.791293653   0.70172726   0.675372459   0   0
#>    0.773610903   0.37375374   0.750912817   0   0
#>    0.692888484   0.57762160   0.864839083   0   0
#>    0.673470158   0.70172726   0.675372459   0   0
#>    0.793813077   0.34594083   0.787105390   0   0
#>    0.865938722   0.27556645   0.903815826   0   0
#>    0.881449679   0.87530919   0.824406608   0   0
#>    0.747919669   0.71823306   0.791293653   0   0
#>    0.905361479   0.70928376   0.849980525   0   0
#>    0.689679217   0.52967320   0.692888484   0   0
#>    0.734077232   0.87530919   0.824406608   0   0
#>    0.683178854   0.79133524   0.651955555   0   0
#>    0.579194795   0.81088575   0.574042900   0   0
#>    0.668779500   0.30131261   0.681099888   0   0
#>    0.959717158   0.27556645   0.903815826   0   0
#>    0.637383133   0.49606429   0.660895647   0   0
#>    0.769009158   0.16686955   0.768638619   0   0
#>    9.334422609   0.94391639   0.898780823   0   0
#>    0.898780823   3.02003239   0.726866429   0   0
#>    0.772381635   0.16911133   0.769009158   0   0
#>    1.007626615  -1.77941858   6.413254851   0   1
#>    0.569605740   0.40213405   0.569585719   0   0
#>    0.644990122   0.39169614   0.569605740   0   0
#>    0.681099888   0.19207104   0.613392402   0   0
#>    0.656391928   0.32690310   0.668779500   0   0
#>    0.490945243   0.45361162   0.476317130   0   0
#>    0.505973576   0.61713511   0.456293804   0   0
#>    0.568177905   0.83764194   0.579194795   0   0
#>    0.607248379   0.32832357   0.548144872   0   0
#>    0.719838019   3.02003239   0.726866429   0   0
#>    0.519876285   0.24324542   0.509618550   0   0
#>    0.613392402   0.24228854   0.593077187   0   0
#>    0.456293804   0.62308496   0.324789814   0   0
#>    0.524225237   0.66275113   0.505973576   0   0
#>    0.513931813   0.25965898   0.519876285   0   0
#>    0.495983715   0.16997128   0.608985989   0   0
#>    0.768638619   0.17616416   0.616003761   0   0
#>    0.466879570   0.35108442   0.459710820   0   0
#>    0.434741509   0.27873949   0.425191849   0   0
#>    0.534302372   0.10555856   0.592816872   0   0
#>    0.424738695   1.03050513   0.445605579   0   0
#>    0.479520373   0.80613648   0.448315981   0   0
#>    0.303200670   0.55955171   0.280623658   0   0
#>    0.473568204   0.27190296   0.430976981   0   0
#>    0.639715511 -11.06491558   2.478331823   0   1
#>    0.803885164  -2.54342661   3.244609675   0   1
#>    0.316457441   0.34286560   0.317167669   0   0
#>    0.339107520   0.33259930   0.434741509   0   0
#>    0.592816872   0.07052011   0.597521970   0   0
#>    0.599668033   0.07052011   0.597521970   0   0
#>    0.495797356  -4.80125237   0.581493486   0   1
#>    0.448581476   0.90406596   0.568177905   0   0
#>    0.280623658   0.62308496   0.324789814   0   0
#>    0.288712999   0.74645604   0.307831612   0   0
#>    0.325002158   0.18622212   0.189341838   0   0
#>    0.465117177  -0.01315949   0.495797356   0   0
#>    0.279673878   0.35638665   0.276179203   0   0
#>    0.318474479   0.27029034   0.325002158   0   0
#>    0.445605579   1.03700549   0.448581476   0   0
#>    0.307831612   0.80613648   0.448315981   0   0
#>    0.225641504   0.53059329   0.212696374   0   0
#>    0.273874555   0.35815830   0.279673878   0   0
#>    0.204288880   0.36592362   0.192821690   0   0
#>    0.188316676   0.23512241   0.184642122   0   0
#>    0.191920103   0.35638665   0.276179203   0   0
#>    0.189341838   0.18860664   0.184944959   0   0
#>    0.360862396   0.13439357   0.405707677   0   0
#>    0.289759742   0.95533067   0.279367255   0   0
#>    0.182604528   0.84794795   0.090407153   0   0
#>    0.192821690   0.40740775   0.166655706   0   0
#>    0.277192523   0.95173764   0.161800850   0   0
#>    0.118186002   0.65903126   0.108544980   0   0
#>    0.293126769   0.04102563   0.461895275   0   0
#>    0.461895275   0.03740139   0.465117177   0   0
#>    0.184642122   0.21061431   0.177189539   0   0
#>    0.199424975   0.84630009   0.182604528   0   0
#>    0.100327926   0.39919482   0.119081627   0   0
#>    0.174812167   0.33708325   0.172010524   0   0
#>    0.055429050   0.77409838   0.044702486   0   0
#>    0.166976398   0.95173764   0.161800850   0   0
#>    0.016939094   0.77409838   0.044702486   0   0
#>    0.184944959   0.21061431   0.177189539   0   0
#>    0.066983911   0.33020586  -0.064649650   0   0
#>    0.018087241   0.36982460   0.066983911   0   0
#>   -0.005024612   0.54078911   0.018087241   0   0
#>    0.090407153   0.86206835   0.089209312   0   0
#>    0.070446949   0.78891738   0.055429050   0   0
#>   -0.077207821   0.87610497   0.021867225   0   0
#>    0.021867225   0.77733072  -0.076170294   0   0
#>    0.065643605   0.17872542   0.025549402   0   0
#>    0.136656438   0.06072773   0.115188708   0   0
#>   -0.707340936   0.43534121  -3.039305394   0   0
#>    0.017267047   0.23917251   0.034935047   0   0
#>    0.052175417   0.17872542   0.025549402   0   0
#>    0.168233469   0.06072773   0.115188708   0   0
#>   -3.039305394   0.55709781  -0.005024612   0   0
#>    1.001356849   0.51871574   0.939313202   0   0
#>    0.968657120   0.40491944   0.874926739   0   0
#>    0.730303443   0.51674071   0.688621261   0   0
#>    0.912352153   0.46568990   1.030161373   0   0
#>    0.793813077   0.31205991   0.865938722   0   0
#>    0.734077232   0.80049732   0.683178854   0   0
#>    9.334422609   1.57014624  11.905383797   0   1
#>    0.644990122   0.33181532   0.656391928   0   0
#>    0.719838019   1.10492805   0.696863945   0   0
#>    0.495983715   0.17145199   0.473568204   0   0
#>    0.782679510   0.06929423   0.599668033   0   0
#>    0.356706531   0.50981940   0.303200670   0   0
#>    0.424738695   0.81002857   0.479520373   0   0
#>    0.470908283   0.13116410   0.418687439   0   0
#>    0.301214630   0.88043028   0.289759742   0   0
#>    0.273874555   0.44017631   0.199523279   0   0
#>    0.188316676   0.28928376   0.174812167   0   0
#>    0.360862396   0.14658831   0.237476777   0   0
#>    0.277192523  11.40968053  -2.723846259   0   1
#>    0.293126769  -0.02647384   0.259719031   0   0
#>    0.024225771   0.77093516   0.016939094   0   0
#>    0.199424975   0.93515158   0.166976398   0   0
#>   -0.077207821   4.99778491  -1.774121608   0   1
#>   -0.064649650   0.27784845   0.017267047   0   0
#>    0.136656438   0.14379038   0.052175417   0   0
#>  -10.371718208   0.48248220 -11.741912742   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.640441321 0.906091940
#>   [2,] 0.452004759 0.148213379
#>   [3,] 0.292249979 0.847325267
#>   [4,] 0.310649092 0.739408802
#>   [5,] 0.426926355 0.015664019
#>   [6,] 0.018260031 0.687602408
#>   [7,] 0.972668565 0.971972912
#>   [8,] 0.870622338 0.000000000
#>   [9,] 0.859662118 0.387574574
#>  [10,] 0.003736775 0.599920661
#>  [11,] 0.790875433 0.814901017
#>  [12,] 0.581236117 0.124551505
#>  [13,] 0.333939191 0.148769619
#>  [14,] 0.742636369 0.932293199
#>  [15,] 0.249398283 0.047686234
#>  [16,] 0.043804776 0.380150328
#>  [17,] 0.388830962 0.211196877
#>  [18,] 0.222851881 0.034042737
#>  [19,] 0.379253737 0.537947083
#>  [20,] 0.207015697 0.891825850
#>  [21,] 0.226396680 0.425580679
#>  [22,] 0.553216947 0.573473036
#>  [23,] 0.910230911 0.509134729
#>  [24,] 0.335310628 0.206672818
#>  [25,] 0.812849365 0.239856564
#>  [26,] 0.244144071 0.467437723
#>  [27,] 0.362577812 0.587190021
#>  [28,] 0.000000000 0.391136120
#>  [29,] 0.226964112 0.296649156
#>  [30,] 0.454544086 0.875618479
#>  [31,] 0.218975721 0.887475286
#>  [32,] 0.369038491 0.952830732
#>  [33,] 1.000000000 0.313048880
#>  [34,] 0.401334126 0.924426530
#>  [35,] 0.395971729 0.690738900
#>  [36,] 0.097225613 0.072531109
#>  [37,] 0.853595941 0.742260701
#>  [38,] 0.331397344 0.885042588
#>  [39,] 0.880191344 0.044037843
#>  [40,] 0.308777572 0.126934063
#>  [41,] 0.901881312 0.117206325
#>  [42,] 0.391140581 0.386841572
#>  [43,] 0.643086998 0.186352907
#>  [44,] 0.879025274 0.671200181
#>  [45,] 0.936090170 0.226835318
#>  [46,] 0.922934924 0.974567020
#>  [47,] 0.527389794 0.808938454
#>  [48,] 0.606601943 0.960550784
#>  [49,] 0.854197313 0.016548455
#>  [50,] 0.359208741 0.164233162
#>  [51,] 0.591678714 0.791011495
#>  [52,] 0.277815730 0.560262400
#>  [53,] 0.336675803 0.195459590
#>  [54,] 0.073090459 0.169951187
#>  [55,] 0.264183085 0.273584690
#>  [56,] 0.102377463 0.656266127
#>  [57,] 0.236228625 0.018857338
#>  [58,] 0.963430047 0.568269173
#>  [59,] 0.622109019 0.953009678
#>  [60,] 0.389009658 0.827916501
#>  [61,] 0.739330439 0.402583430
#>  [62,] 0.140801819 0.001443768
#>  [63,] 0.974618732 0.224656090
#>  [64,] 0.379375279 0.601311898
#>  [65,] 0.687757834 0.721254497
#>  [66,] 0.727947963 0.715531901
#>  [67,] 0.079929295 0.664357598
#>  [68,] 0.645635090 0.687824029
#>  [69,] 0.092554032 0.321393023
#>  [70,] 0.951119352 0.335153149
#>  [71,] 0.261555328 0.637436881
#>  [72,] 0.500289322 0.391768101
#>  [73,] 0.085751568 0.532454546
#>  [74,] 0.106492324 0.510926607
#>  [75,] 0.819600039 0.368483373
#>  [76,] 0.056949761 0.059175380
#>  [77,] 0.244520269 0.664921825
#>  [78,] 0.293882672 0.055707170
#>  [79,] 0.436872847 0.252075207
#>  [80,] 0.416984473 0.798301035
#>  [81,] 0.406841965 0.363130727
#>  [82,] 0.441099207 0.572081348
#>  [83,] 0.191914383 0.068506735
#>  [84,] 0.225292721 0.547798783
#>  [85,] 0.689719117 0.035288251
#>  [86,] 0.422131876 0.318882213
#>  [87,] 0.887541456 0.130682502
#>  [88,] 0.132720865 0.535387468
#>  [89,] 0.806940503 0.129241906
#>  [90,] 0.585739110 1.000000000
#>  [91,] 0.284793899 0.263939325
#>  [92,] 0.688508422 0.005167514
#>  [93,] 0.445426809 0.080349835
#>  [94,] 0.704947950 0.627519212
#>  [95,] 0.327312973 0.843438986
#>  [96,] 0.960238047 0.821856360
#>  [97,] 0.772375036 0.129587811
#>  [98,] 0.287972482 0.373668329
#>  [99,] 0.269695689 0.476748912
#> [100,] 0.928961922 0.622069409
#> 
#> $tri.obj
#> Delauney triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,59,48), [2,4,8]
#> 2: (90,48,59), [1,7,5]
#> 3: (47,51,1), [14,4,26]
#> 4: (47,1,48), [1,6,3]
#> 5: (48,90,30), [10,6,2]
#> 6: (48,30,47), [12,4,5]
#> 7: (14,90,59), [2,8,37]
#> 8: (14,59,1), [1,16,7]
#> 9: (11,1,65), [14,21,16]
#> 10: (34,30,90), [5,19,17]
#> 11: (35,47,80), [12,27,20]
#> 12: (80,47,30), [6,13,11]
#> 13: (30,60,80), [25,12,17]
#> 14: (1,51,65), [15,9,3]
#> 15: (51,68,65), [29,14,26]
#> 16: (11,14,1), [8,9,38]
#> 17: (30,34,60), [24,13,10]
#> 18: (34,32,38), [46,24,19]
#> 19: (32,34,90), [10,0,18]
#> 20: (47,35,22), [40,39,11]
#> 21: (66,11,65), [9,22,35]
#> 22: (65,94,66), [28,21,29]
#> 23: (60,38,95), [32,36,24]
#> 24: (34,38,60), [23,17,18]
#> 25: (80,60,4), [36,27,13]
#> 26: (47,68,51), [15,3,39]
#> 27: (80,4,35), [44,11,25]
#> 28: (44,66,94), [22,69,42]
#> 29: (65,68,94), [30,22,15]
#> 30: (68,22,94), [62,29,39]
#> 31: (95,3,4), [45,36,32]
#> 32: (3,95,38), [23,33,31]
#> 33: (38,31,3), [45,32,46]
#> 34: (96,46,11), [38,47,51]
#> 35: (66,37,11), [47,21,42]
#> 36: (4,60,95), [23,31,25]
#> 37: (46,90,14), [7,38,50]
#> 38: (11,46,14), [37,16,34]
#> 39: (47,22,68), [30,26,20]
#> 40: (35,82,22), [58,20,48]
#> 41: (37,44,96), [61,47,42]
#> 42: (44,37,66), [35,28,41]
#> 43: (23,100,44), [61,69,63]
#> 44: (71,35,4), [27,56,57]
#> 45: (31,4,3), [31,33,52]
#> 46: (38,32,31), [53,33,18]
#> 47: (96,11,37), [35,41,34]
#> 48: (35,64,82), [59,40,57]
#> 49: (77,31,20), [53,78,52]
#> 50: (7,90,46), [37,51,0]
#> 51: (7,46,96), [34,116,50]
#> 52: (77,4,31), [45,49,56]
#> 53: (20,31,32), [46,0,49]
#> 54: (64,27,19), [60,59,55]
#> 55: (64,71,27), [64,54,57]
#> 56: (71,4,77), [52,71,44]
#> 57: (71,64,35), [48,44,55]
#> 58: (72,22,82), [40,68,73]
#> 59: (64,19,82), [68,48,54]
#> 60: (27,52,19), [76,54,64]
#> 61: (100,96,44), [41,43,67]
#> 62: (61,94,22), [30,73,75]
#> 63: (23,58,100), [67,43,105]
#> 64: (27,71,52), [65,60,55]
#> 65: (71,84,52), [70,64,71]
#> 66: (96,58,33), [105,116,67]
#> 67: (58,96,100), [61,63,66]
#> 68: (72,82,19), [59,79,58]
#> 69: (23,44,94), [28,75,43]
#> 70: (99,52,84), [65,81,76]
#> 71: (71,77,84), [72,65,56]
#> 72: (77,56,84), [92,71,78]
#> 73: (61,22,72), [58,96,62]
#> 74: (42,19,99), [76,80,79]
#> 75: (61,23,94), [69,62,88]
#> 76: (99,19,52), [60,70,74]
#> 77: (88,26,84), [81,92,91]
#> 78: (77,20,56), [83,72,49]
#> 79: (42,72,19), [68,74,85]
#> 80: (98,42,99), [74,86,98]
#> 81: (84,26,99), [93,70,77]
#> 82: (73,74,88), [100,101,113]
#> 83: (67,56,20), [78,95,102]
#> 84: (86,72,81), [85,97,90]
#> 85: (81,72,42), [79,98,84]
#> 86: (98,99,21), [93,112,80]
#> 87: (23,9,70), [106,117,88]
#> 88: (9,23,61), [75,89,87]
#> 89: (61,75,9), [106,88,119]
#> 90: (72,86,79), [121,107,84]
#> 91: (26,88,21), [100,93,77]
#> 92: (88,84,56), [72,101,77]
#> 93: (21,99,26), [81,91,86]
#> 94: (10,67,6), [95,0,102]
#> 95: (6,67,20), [83,0,94]
#> 96: (72,43,61), [108,73,107]
#> 97: (91,86,81), [84,109,114]
#> 98: (98,81,42), [85,80,109]
#> 99: (69,21,16), [129,141,128]
#> 100: (74,21,88), [91,82,129]
#> 101: (73,88,56), [92,104,82]
#> 102: (56,67,10), [94,104,83]
#> 103: (28,73,10), [104,0,113]
#> 104: (56,10,73), [103,101,102]
#> 105: (23,33,58), [66,63,117]
#> 106: (70,9,75), [89,111,87]
#> 107: (72,79,43), [120,96,90]
#> 108: (61,43,25), [138,119,96]
#> 109: (91,81,98), [98,115,97]
#> 110: (55,98,29), [112,127,115]
#> 111: (25,70,75), [106,119,130]
#> 112: (29,98,21), [86,128,110]
#> 113: (28,74,73), [82,103,142]
#> 114: (17,86,91), [97,122,121]
#> 115: (55,91,98), [109,110,145]
#> 116: (96,33,7), [0,51,66]
#> 117: (23,70,33), [118,105,87]
#> 118: (70,45,33), [137,117,130]
#> 119: (61,25,75), [111,89,108]
#> 120: (12,43,79), [107,133,149]
#> 121: (17,79,86), [90,114,126]
#> 122: (91,24,17), [123,114,125]
#> 123: (17,24,53), [125,134,122]
#> 124: (40,53,91), [125,145,151]
#> 125: (91,53,24), [123,122,124]
#> 126: (79,17,2), [135,133,121]
#> 127: (55,29,54), [132,158,110]
#> 128: (69,29,21), [112,99,132]
#> 129: (21,74,16), [142,99,100]
#> 130: (70,25,45), [148,118,111]
#> 131: (87,25,89), [143,166,148]
#> 132: (29,69,54), [159,127,128]
#> 133: (12,79,2), [126,154,120]
#> 134: (17,53,50), [136,135,123]
#> 135: (17,50,2), [144,126,134]
#> 136: (53,13,50), [150,134,151]
#> 137: (63,33,45), [118,146,0]
#> 138: (97,25,43), [108,139,143]
#> 139: (85,97,43), [138,149,162]
#> 140: (13,40,93), [161,150,151]
#> 141: (28,69,16), [99,142,159]
#> 142: (28,16,74), [129,113,141]
#> 143: (89,25,97), [138,153,131]
#> 144: (50,93,2), [154,135,150]
#> 145: (40,91,55), [115,152,124]
#> 146: (45,41,63), [165,137,155]
#> 147: (92,85,12), [149,164,157]
#> 148: (87,45,25), [130,131,155]
#> 149: (85,43,12), [120,147,139]
#> 150: (50,13,93), [140,144,136]
#> 151: (53,40,13), [140,136,124]
#> 152: (55,83,40), [170,145,158]
#> 153: (49,89,97), [143,162,169]
#> 154: (93,12,2), [133,144,163]
#> 155: (45,87,41), [156,146,148]
#> 156: (87,39,41), [165,155,166]
#> 157: (85,92,49), [173,162,147]
#> 158: (55,54,83), [176,152,127]
#> 159: (69,28,54), [183,132,141]
#> 160: (5,93,78), [161,174,163]
#> 161: (40,78,93), [160,140,168]
#> 162: (49,97,85), [139,157,153]
#> 163: (5,12,93), [154,160,164]
#> 164: (5,92,12), [147,163,185]
#> 165: (39,63,41), [146,156,171]
#> 166: (87,89,39), [169,156,131]
#> 167: (57,15,18), [175,179,178]
#> 168: (15,78,40), [161,170,178]
#> 169: (49,39,89), [166,153,172]
#> 170: (83,15,40), [168,152,175]
#> 171: (8,63,39), [165,172,0]
#> 172: (8,39,49), [169,173,171]
#> 173: (8,49,92), [157,184,172]
#> 174: (78,57,5), [177,160,178]
#> 175: (18,15,83), [170,186,167]
#> 176: (36,83,54), [158,181,180]
#> 177: (5,57,62), [179,185,174]
#> 178: (57,78,15), [168,167,174]
#> 179: (62,57,18), [167,186,177]
#> 180: (83,36,62), [182,186,176]
#> 181: (54,76,36), [182,176,183]
#> 182: (36,76,62), [0,180,181]
#> 183: (54,28,76), [0,181,159]
#> 184: (62,8,92), [173,185,0]
#> 185: (5,62,92), [184,164,177]
#> 186: (83,62,18), [179,175,180]
#> boundary nodes:  90 32 20 6 10 28 76 62 8 63 33 7 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                              
#> 0.01724348 0.12498520 0.06411606 0.18078412 0.09875587 0.05942105 0.05037209 
#>                                                                              
#> 0.19079762 0.11091659 0.04073900 0.11873178 0.05377659 0.10550040 0.09729377 
#>                                                                              
#> 0.07753766 0.19929011 0.19615642 0.13929433 0.09096831 0.04180361 0.08113303 
#>                                                                              
#> 0.11821164 0.11644288 0.09822786 0.08458538 0.16106920 0.10947370 0.08355322 
#>                                                                              
#> 0.20720266 0.09597123 0.06361940 0.17088289 0.18518813 0.14702810 0.11212638 
#>                                                                              
#> 0.17115139 0.12845950 0.07005356 0.16367703 0.06829545 0.01272673 0.33815359 
#>                                                                              
#> 0.15720073 0.17414847 0.05198994 0.11282858 0.09960588 0.09095400 0.07064000 
#>                                                                              
#> 0.10386358 0.06389458 0.07886889 0.05398155 0.25782699 0.20222023 0.17947407 
#>                                                                              
#> 0.11869080 0.16393086 0.18925648 0.12549176 0.23697253 0.08254255 0.25775481 
#>                                                                              
#> 0.18978384 0.15110325 0.02719528 0.03586461 0.09773693 0.10925987 0.06702979 
#>                                                                              
#> 0.10541552 0.20139551 0.04437842 0.06841398 0.14433534 0.06590462 0.26056239 
#>                                                                              
#> 0.23668215 0.04681573 0.02843831 0.18815876 0.12462894 0.09978687 0.10619841 
#>                                                                              
#> 0.21643405 0.17799411 0.11933564 0.09825464 0.13567695 0.08053863 0.02989366 
#>                                                                              
#> 0.14792039 0.10977503 0.65949063 0.05364622 0.10731843 0.12880383 0.05371120 
#>                                                                              
#> 0.01129602 0.08591395 0.08920678 0.19932254 0.12893277 0.14502470 0.12392679 
#>                                                                              
#> 0.11077237 0.15268746 0.10495834 0.03850744 0.09416865 0.02962552 0.14120083 
#>                                                                              
#> 0.14437141 0.07634730 0.11746220 0.06818160 0.02275607 0.12975432 0.14048662 
#>                                                                              
#> 0.08743538 0.13080809 0.03331514 0.13065511 0.03456720 0.13137971 0.01967838 
#>                                                                              
#> 0.07631567 0.16607930 0.15623732 0.23294847 0.15353464 0.12539700 0.14282141 
#>                                                                              
#> 0.16049886 0.11236289 0.02984722 0.07276763 0.09902585 0.20381237 0.03783327 
#>                                                                              
#> 0.19072447 0.06113830 0.09700262 0.04520173 0.02023694 0.08338043 0.04243253 
#>                                                                              
#> 0.10180426 0.33681034 0.18218722 0.54771926 0.08828876 0.14906189 0.17105807 
#>                                                                              
#> 0.17416971 0.12229479 0.10959538 0.08595762 0.17591534 0.08105702 0.08026362 
#>                                                                              
#> 0.22177488 0.23687762 0.04059550 0.12155455 0.06674160 0.09529850 0.10535637 
#>                                                                              
#> 0.05436095 0.16950537 0.11761840 0.12691703 0.16926504 0.13307144 0.15744731 
#>                                                                              
#> 0.16504243 0.11317451 0.11244794 0.12694922 0.22998267 0.04980125 0.15063033 
#>                                                                              
#> 0.22401466 0.17312707 0.06336493 0.02194498 0.03233598 0.14459901 0.18977971 
#>                                                                              
#> 0.25266873 0.11447748 0.08893654 0.51035876 0.08378370 0.09669522 0.23928566 
#>                                                                              
#> 0.22754837 0.08390732 0.09340016 0.14240610 0.15157233 0.10468834 0.04706078 
#>                                                                              
#> 0.17872172 0.12126409 0.15341367 0.04546407 0.08887640 0.27809336 0.15727258 
#>                                                                              
#> 0.10400573 0.12492289 0.11359928 0.20881798 0.07954263 0.25017310 0.17856390 
#>                                                                              
#> 0.04378609 0.16530042 0.11664251 0.10287211 0.21565748 0.08721251 0.19262152 
#>                                                                              
#> 0.11271682 0.05447773 0.13908867 0.07636356 0.21738513 0.16961404 0.10935552 
#>                                                                              
#> 0.08061383 0.05552537 0.03859014 0.15809834 0.04516133 0.16028218 0.15328103 
#>                                                                              
#> 0.13227957 0.12029118 0.04677010 0.13954491 0.11484234 0.16554231 0.21743903 
#>                                                                              
#> 0.13893911 0.06727943 0.18885970 0.08695586 0.12220074 0.04506548 0.02331593 
#>                                                                              
#> 0.04631302 0.10036522 0.28647769 0.03169456 0.08432043 0.11194553 0.26179258 
#>                                                                              
#> 0.04462621 0.07220475 0.17675362 0.04300938 0.03527771 0.07547350 0.38794319 
#>                                                                              
#> 0.12323374 0.25360728 0.13050785 0.02386194 0.10686879 0.13165893 0.14717670 
#>                                                                              
#> 0.16792864 0.06307980 0.07398689 0.13666868 0.09196463 0.11588938 0.03014506 
#>                                                        
#> 0.10771417 0.24755930 0.06842441 0.09477425 0.72982195 
#> 
#> $alpha
#> [1] 0.2881995
#> 
#> attr(,"class")
#> [1] "scree"