Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, global_rng = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

global_rng

an optional vector c(min, max) that specifies the global range of the data set if you want all variables one the same scale

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2          mx1
#>    73   18 0.051832779 0.737785622 0.015059471 0.645633270  0.100955105
#>    18   23 0.015059471 0.645633270 0.143448566 0.545763542  0.097849311
#>    23   87 0.143448566 0.545763542 0.178527722 0.572889854  0.123279848
#>    87    8 0.178527722 0.572889854 0.191193699 0.608079255  0.139569329
#>     8   90 0.191193699 0.608079255 0.290916500 0.743106880  0.278106037
#>    90   86 0.290916500 0.743106880 0.223869562 0.735728165  0.266129276
#>    86   73 0.223869562 0.735728165 0.051832779 0.737785622  0.137545381
#>    90   72 0.290916500 0.743106880 0.303688235 0.480078236  0.329733221
#>    72   89 0.303688235 0.480078236 0.385756391 0.427269552  0.376359062
#>    89   21 0.385756391 0.427269552 0.599446304 0.328526417  0.516899598
#>    21   79 0.599446304 0.328526417 0.619077113 0.527516275  0.524817083
#>    79   62 0.619077113 0.527516275 0.480116162 0.571428754  0.527704827
#>    62   67 0.480116162 0.571428754 0.459630210 0.569024062  0.473156916
#>    67   90 0.459630210 0.569024062 0.290916500 0.743106880  0.335304633
#>    48   61 0.628619014 0.628089405 0.778967292 0.516481771  0.723903555
#>    61   81 0.778967292 0.516481771 0.768613429 0.705857969  0.742295398
#>    81   71 0.768613429 0.705857969 0.739585924 0.816128372  0.716534305
#>    71   30 0.739585924 0.816128372 0.573853630 0.770283473  0.661614406
#>    30   48 0.573853630 0.770283473 0.628619014 0.628089405  0.608035550
#>    86   25 0.223869562 0.735728165 0.215070571 0.886024854  0.187656370
#>    25   19 0.215070571 0.886024854 0.103054608 0.858497681  0.165859250
#>    19   52 0.103054608 0.858497681 0.065019084 0.836033201  0.089705985
#>    52   73 0.065019084 0.836033201 0.051832779 0.737785622  0.069069199
#>    99   40 0.804828527 0.948397027 0.619035173 0.995902369  0.720405396
#>    40   60 0.619035173 0.995902369 0.612647717 0.941532274  0.655397036
#>    60   32 0.612647717 0.941532274 0.574799206 0.827106327  0.617923634
#>    32   71 0.574799206 0.827106327 0.739585924 0.816128372  0.658062488
#>    71   99 0.739585924 0.816128372 0.804828527 0.948397027  0.745219090
#>    76   95 0.728527356 0.237191695 0.716175906 0.203998527  0.747740923
#>    95   41 0.716175906 0.203998527 0.688585342 0.147539926  0.750846423
#>    41   13 0.688585342 0.147539926 0.863671572 0.076393412  0.776511001
#>    13   80 0.863671572 0.076393412 0.900826439 0.259193534  0.867102487
#>    80   84 0.900826439 0.259193534 0.810251584 0.251320595  0.861798401
#>    84   39 0.810251584 0.251320595 0.780383685 0.266639149  0.784024505
#>    39   76 0.780383685 0.266639149 0.728527356 0.237191695  0.762845965
#>    23   14 0.143448566 0.545763542 0.146685177 0.440101403  0.197217398
#>    14   46 0.146685177 0.440101403 0.163858993 0.326911882  0.200258527
#>    46   35 0.163858993 0.326911882 0.261125762 0.346990339  0.201694983
#>    35   93 0.261125762 0.346990339 0.259178490 0.436899280  0.218351434
#>    93   72 0.259178490 0.436899280 0.303688235 0.480078236  0.274318642
#>    72   87 0.303688235 0.480078236 0.178527722 0.572889854  0.236379268
#>    35   69 0.261125762 0.346990339 0.195210262 0.287009488  0.282730095
#>    69   38 0.195210262 0.287009488 0.165089772 0.175707343  0.246939462
#>    38   65 0.165089772 0.175707343 0.162467003 0.000000000  0.248089238
#>    65   47 0.162467003 0.000000000 0.712209092 0.008312603  0.438809218
#>    47   34 0.712209092 0.008312603 0.590724609 0.181206863  0.463413426
#>    34   78 0.590724609 0.181206863 0.488940117 0.229495902  0.462200152
#>    78   74 0.488940117 0.229495902 0.378467486 0.213576565  0.457590410
#>    74   33 0.378467486 0.213576565 0.389482030 0.262464955  0.309937115
#>    33   35 0.389482030 0.262464955 0.261125762 0.346990339  0.297927795
#>    56   24 0.606792692 0.293291532 0.691244639 0.368943855  0.668530832
#>    24   21 0.691244639 0.368943855 0.599446304 0.328526417  0.645143693
#>    21   56 0.599446304 0.328526417 0.606792692 0.293291532  0.569787526
#>    18   91 0.015059471 0.645633270 0.000000000 0.616845442  0.072008296
#>    91   58 0.000000000 0.616845442 0.001088476 0.573235405  0.071397456
#>    58   36 0.001088476 0.573235405 0.075912801 0.403631229  0.019444151
#>    36   14 0.075912801 0.403631229 0.146685177 0.440101403  0.088444886
#>    84   43 0.810251584 0.251320595 0.808624172 0.285036254  0.805008373
#>    43   17 0.808624172 0.285036254 0.834602960 0.376339195  0.771952283
#>    17   24 0.834602960 0.376339195 0.691244639 0.368943855  0.764179450
#>    24   39 0.691244639 0.368943855 0.780383685 0.266639149  0.735930987
#>    88   75 0.853966452 0.043819248 0.978491797 0.305643237  1.657666684
#>    75   80 0.978491797 0.305643237 0.900826439 0.259193534  1.025644756
#>    41   47 0.688585342 0.147539926 0.712209092 0.008312603  0.766879995
#>    47   88 0.712209092 0.008312603 0.853966452 0.043819248  0.767766608
#>    36   46 0.075912801 0.403631229 0.163858993 0.326911882  0.132821692
#>    58   70 0.001088476 0.573235405 0.030713736 0.243289779 -0.269770989
#>    70   26 0.030713736 0.243289779 0.126459136 0.256474787  0.072900501
#>    26   46 0.126459136 0.256474787 0.163858993 0.326911882  0.104969491
#>    68   63 0.563175078 0.969397159 0.227194157 0.953475893  0.377790411
#>    63   97 0.227194157 0.953475893 0.277619210 0.913720588  0.289561258
#>    97   90 0.277619210 0.913720588 0.290916500 0.743106880  0.326471769
#>    90   32 0.290916500 0.743106880 0.574799206 0.827106327  0.425520216
#>    32   68 0.574799206 0.827106327 0.563175078 0.969397159  0.458072712
#>    43   28 0.808624172 0.285036254 0.854752695 0.329243854  0.804437771
#>    28   31 0.854752695 0.329243854 0.858916565 0.334011152  0.836573472
#>    31   75 0.858916565 0.334011152 0.978491797 0.305643237  0.939763996
#>    75   92 0.978491797 0.305643237 0.898345609 0.692717509  0.989462152
#>    92   61 0.898345609 0.692717509 0.778967292 0.516481771  0.974118343
#>    61   50 0.778967292 0.516481771 0.744746717 0.475157316  0.816747108
#>    50   24 0.744746717 0.475157316 0.691244639 0.368943855  0.761505578
#>    73   51 0.051832779 0.737785622 0.005778683 0.905174578 -0.160933910
#>    51   91 0.005778683 0.905174578 0.000000000 0.616845442 -0.250239033
#>    66   52 0.094688224 0.874492507 0.065019084 0.836033201  0.078861267
#>    25   66 0.215070571 0.886024854 0.094688224 0.874492507  0.153462432
#>    42   44 0.110262862 0.984665025 0.033344137 1.000000000  0.069973759
#>    44   57 0.033344137 1.000000000 0.061486149 0.936628334  0.053342267
#>    57   53 0.061486149 0.936628334 0.071509290 0.906151205  0.075718287
#>    53   66 0.071509290 0.906151205 0.094688224 0.874492507  0.123411293
#>    66   42 0.094688224 0.874492507 0.110262862 0.984665025  0.139291611
#>    24   82 0.691244639 0.368943855 0.672558887 0.468233633  0.619740587
#>    82   79 0.672558887 0.468233633 0.619077113 0.527516275  0.572416844
#>    89   78 0.385756391 0.427269552 0.488940117 0.229495902  0.480004777
#>    78   56 0.488940117 0.229495902 0.606792692 0.293291532  0.529384878
#>    70   85 0.030713736 0.243289779 0.072054285 0.136988743  0.009415969
#>    85   55 0.072054285 0.136988743 0.098750497 0.213936478  0.098903322
#>    55   70 0.098750497 0.213936478 0.030713736 0.243289779  0.074034905
#>    56   76 0.606792692 0.293291532 0.728527356 0.237191695  0.681373525
#>    50   29 0.744746717 0.475157316 0.695166069 0.458187902  0.738818618
#>    29   82 0.695166069 0.458187902 0.672558887 0.468233633  0.662403683
#>    63   42 0.227194157 0.953475893 0.110262862 0.984665025  0.158572576
#>    55   38 0.098750497 0.213936478 0.165089772 0.175707343  0.143568955
#>    69   46 0.195210262 0.287009488 0.163858993 0.326911882  0.154076651
#>    45   51 0.007611850 0.990496885 0.005778683 0.905174578 -0.052660132
#>    51   77 0.005778683 0.905174578 0.057692616 0.907970747  0.031367310
#>    77   57 0.057692616 0.907970747 0.061486149 0.936628334  0.040241463
#>    44   45 0.033344137 1.000000000 0.007611850 0.990496885  0.030594731
#>    80   96 0.900826439 0.259193534 0.914527284 0.287351136  0.872193236
#>    96   31 0.914527284 0.287351136 0.858916565 0.334011152  0.884224452
#>    43   80 0.808624172 0.285036254 0.900826439 0.259193534  0.857108004
#>    50   82 0.744746717 0.475157316 0.672558887 0.468233633  0.704921207
#>    90   64 0.290916500 0.743106880 0.529637143 0.687820573  0.428703223
#>    64   48 0.529637143 0.687820573 0.628619014 0.628089405  0.604849106
#>    96   75 0.914527284 0.287351136 0.978491797 0.305643237  0.928286763
#>    60   68 0.612647717 0.941532274 0.563175078 0.969397159  0.555045364
#>    89   98 0.385756391 0.427269552 0.349568485 0.426105697  0.370306486
#>    98   83 0.349568485 0.426105697 0.275156591 0.364444562  0.357121940
#>    83   35 0.275156591 0.364444562 0.261125762 0.346990339  0.327473501
#>    95   56 0.716175906 0.203998527 0.606792692 0.293291532  0.646584609
#>    83   93 0.275156591 0.364444562 0.259178490 0.436899280  0.224047430
#>    75   49 0.978491797 0.305643237 1.000000000 0.852088542  1.294480586
#>    49  100 1.000000000 0.852088542 0.918530744 0.973604099  0.897285657
#>   100   99 0.918530744 0.973604099 0.804828527 0.948397027  0.881981562
#>    71   59 0.739585924 0.816128372 0.793433620 0.716379705  0.843174219
#>    59   54 0.793433620 0.716379705 0.825040919 0.691683649  0.859283849
#>    54   92 0.825040919 0.691683649 0.898345609 0.692717509  0.860649528
#>    85   65 0.072054285 0.136988743 0.162467003 0.000000000  0.146972949
#>    64   62 0.529637143 0.687820573 0.480116162 0.571428754  0.550355297
#>    79   61 0.619077113 0.527516275 0.778967292 0.516481771  0.702359330
#>    61   82 0.778967292 0.516481771 0.672558887 0.468233633  0.701549100
#>    52   77 0.065019084 0.836033201 0.057692616 0.907970747  0.033749121
#>    66   77 0.094688224 0.874492507 0.057692616 0.907970747  0.058528453
#>    59   81 0.793433620 0.716379705 0.768613429 0.705857969  0.792894642
#>    64   67 0.529637143 0.687820573 0.459630210 0.569024062  0.460694295
#>    93   94 0.259178490 0.436899280 0.331741754 0.424328106  0.297432188
#>    94   98 0.331741754 0.424328106 0.349568485 0.426105697  0.336996694
#>     1   18 0.183422809 0.634055985 0.015059471 0.645633270  0.100955105
#>     1   23 0.183422809 0.634055985 0.143448566 0.545763542  0.097849311
#>     1   87 0.183422809 0.634055985 0.178527722 0.572889854  0.123279848
#>     1    8 0.183422809 0.634055985 0.191193699 0.608079255  0.139569329
#>     1   90 0.183422809 0.634055985 0.290916500 0.743106880  0.278106037
#>     1   86 0.183422809 0.634055985 0.223869562 0.735728165  0.266129276
#>     1   73 0.183422809 0.634055985 0.051832779 0.737785622  0.137545381
#>     2   72 0.443328389 0.539091822 0.303688235 0.480078236  0.329733221
#>     2   89 0.443328389 0.539091822 0.385756391 0.427269552  0.376359062
#>     2   21 0.443328389 0.539091822 0.599446304 0.328526417  0.516899598
#>     2   79 0.443328389 0.539091822 0.619077113 0.527516275  0.524817083
#>     2   62 0.443328389 0.539091822 0.480116162 0.571428754  0.527704827
#>     2   67 0.443328389 0.539091822 0.459630210 0.569024062  0.473156916
#>     2   90 0.443328389 0.539091822 0.290916500 0.743106880  0.335304633
#>     3   61 0.683719461 0.690419492 0.778967292 0.516481771  0.723903555
#>     3   81 0.683719461 0.690419492 0.768613429 0.705857969  0.742295398
#>     3   71 0.683719461 0.690419492 0.739585924 0.816128372  0.716534305
#>     3   30 0.683719461 0.690419492 0.573853630 0.770283473  0.661614406
#>     3   48 0.683719461 0.690419492 0.628619014 0.628089405  0.608035550
#>     4   25 0.106598709 0.819590630 0.215070571 0.886024854  0.187656370
#>     4   19 0.106598709 0.819590630 0.103054608 0.858497681  0.165859250
#>     4   52 0.106598709 0.819590630 0.065019084 0.836033201  0.089705985
#>     4   73 0.106598709 0.819590630 0.051832779 0.737785622  0.069069199
#>     4   86 0.106598709 0.819590630 0.223869562 0.735728165  0.137882927
#>     5   40 0.668680624 0.917605111 0.619035173 0.995902369  0.720405396
#>     5   60 0.668680624 0.917605111 0.612647717 0.941532274  0.655397036
#>     5   32 0.668680624 0.917605111 0.574799206 0.827106327  0.617923634
#>     5   71 0.668680624 0.917605111 0.739585924 0.816128372  0.658062488
#>     5   99 0.668680624 0.917605111 0.804828527 0.948397027  0.745219090
#>     6   95 0.779881603 0.207348143 0.716175906 0.203998527  0.747740923
#>     6   41 0.779881603 0.207348143 0.688585342 0.147539926  0.750846423
#>     6   13 0.779881603 0.207348143 0.863671572 0.076393412  0.776511001
#>     6   80 0.779881603 0.207348143 0.900826439 0.259193534  0.867102487
#>     6   84 0.779881603 0.207348143 0.810251584 0.251320595  0.861798401
#>     6   39 0.779881603 0.207348143 0.780383685 0.266639149  0.784024505
#>     6   76 0.779881603 0.207348143 0.728527356 0.237191695  0.762845965
#>     7   23 0.250838017 0.443141224 0.143448566 0.545763542  0.204887786
#>     7   14 0.250838017 0.443141224 0.146685177 0.440101403  0.197217398
#>     7   46 0.250838017 0.443141224 0.163858993 0.326911882  0.200258527
#>     7   35 0.250838017 0.443141224 0.261125762 0.346990339  0.201694983
#>     7   93 0.250838017 0.443141224 0.259178490 0.436899280  0.218351434
#>     7   72 0.250838017 0.443141224 0.303688235 0.480078236  0.274318642
#>     7   87 0.250838017 0.443141224 0.178527722 0.572889854  0.236379268
#>     8   72 0.191193699 0.608079255 0.303688235 0.480078236  0.266718462
#>     9   69 0.329457673 0.177199079 0.195210262 0.287009488  0.282730095
#>     9   38 0.329457673 0.177199079 0.165089772 0.175707343  0.246939462
#>     9   65 0.329457673 0.177199079 0.162467003 0.000000000  0.248089238
#>     9   47 0.329457673 0.177199079 0.712209092 0.008312603  0.438809218
#>     9   34 0.329457673 0.177199079 0.590724609 0.181206863  0.463413426
#>     9   78 0.329457673 0.177199079 0.488940117 0.229495902  0.462200152
#>     9   74 0.329457673 0.177199079 0.378467486 0.213576565  0.457590410
#>     9   33 0.329457673 0.177199079 0.389482030 0.262464955  0.309937115
#>     9   35 0.329457673 0.177199079 0.261125762 0.346990339  0.297927795
#>    10   24 0.605415617 0.318582045 0.691244639 0.368943855  0.668530832
#>    10   21 0.605415617 0.318582045 0.599446304 0.328526417  0.645143693
#>    10   56 0.605415617 0.318582045 0.606792692 0.293291532  0.569787526
#>    11   91 0.100965396 0.528803004 0.000000000 0.616845442  0.072008296
#>    11   58 0.100965396 0.528803004 0.001088476 0.573235405  0.071397456
#>    11   36 0.100965396 0.528803004 0.075912801 0.403631229  0.019444151
#>    11   14 0.100965396 0.528803004 0.146685177 0.440101403  0.088444886
#>    11   23 0.100965396 0.528803004 0.143448566 0.545763542  0.139975397
#>    11   18 0.100965396 0.528803004 0.015059471 0.645633270  0.092227484
#>    12   43 0.790128630 0.277080977 0.808624172 0.285036254  0.805008373
#>    12   17 0.790128630 0.277080977 0.834602960 0.376339195  0.771952283
#>    12   24 0.790128630 0.277080977 0.691244639 0.368943855  0.764179450
#>    12   39 0.790128630 0.277080977 0.780383685 0.266639149  0.735930987
#>    12   84 0.790128630 0.277080977 0.810251584 0.251320595  0.796528234
#>    13   75 0.863671572 0.076393412 0.978491797 0.305643237  1.657666684
#>    13   47 0.863671572 0.076393412 0.712209092 0.008312603  0.766879995
#>    13   88 0.863671572 0.076393412 0.853966452 0.043819248  0.767766608
#>    15   70 0.060177991 0.353706129 0.030713736 0.243289779 -0.269770989
#>    15   26 0.060177991 0.353706129 0.126459136 0.256474787  0.072900501
#>    15   46 0.060177991 0.353706129 0.163858993 0.326911882  0.104969491
#>    15   36 0.060177991 0.353706129 0.075912801 0.403631229  0.117873341
#>    15   58 0.060177991 0.353706129 0.001088476 0.573235405 -0.094314297
#>    16   63 0.331396052 0.927040345 0.227194157 0.953475893  0.377790411
#>    16   97 0.331396052 0.927040345 0.277619210 0.913720588  0.289561258
#>    16   90 0.331396052 0.927040345 0.290916500 0.743106880  0.326471769
#>    16   32 0.331396052 0.927040345 0.574799206 0.827106327  0.425520216
#>    16   68 0.331396052 0.927040345 0.563175078 0.969397159  0.458072712
#>    17   28 0.834602960 0.376339195 0.854752695 0.329243854  0.804437771
#>    17   31 0.834602960 0.376339195 0.858916565 0.334011152  0.836573472
#>    17   75 0.834602960 0.376339195 0.978491797 0.305643237  0.939763996
#>    17   92 0.834602960 0.376339195 0.898345609 0.692717509  0.989462152
#>    17   61 0.834602960 0.376339195 0.778967292 0.516481771  0.974118343
#>    17   50 0.834602960 0.376339195 0.744746717 0.475157316  0.816747108
#>    18   51 0.015059471 0.645633270 0.005778683 0.905174578 -0.160933910
#>    19   66 0.103054608 0.858497681 0.094688224 0.874492507  0.153462432
#>    20   44 0.075067561 0.943160388 0.033344137 1.000000000  0.069973759
#>    20   57 0.075067561 0.943160388 0.061486149 0.936628334  0.053342267
#>    20   53 0.075067561 0.943160388 0.071509290 0.906151205  0.075718287
#>    20   66 0.075067561 0.943160388 0.094688224 0.874492507  0.123411293
#>    20   42 0.075067561 0.943160388 0.110262862 0.984665025  0.139291611
#>    21   82 0.599446304 0.328526417 0.672558887 0.468233633  0.619740587
#>    21   78 0.599446304 0.328526417 0.488940117 0.229495902  0.480004777
#>    22   85 0.074408717 0.206312247 0.072054285 0.136988743  0.009415969
#>    22   55 0.074408717 0.206312247 0.098750497 0.213936478  0.098903322
#>    22   70 0.074408717 0.206312247 0.030713736 0.243289779  0.074034905
#>    24   76 0.691244639 0.368943855 0.728527356 0.237191695  0.681373525
#>    24   29 0.691244639 0.368943855 0.695166069 0.458187902  0.738818618
#>    25   90 0.215070571 0.886024854 0.290916500 0.743106880  0.249336329
#>    25   97 0.215070571 0.886024854 0.277619210 0.913720588  0.278195566
#>    25   63 0.215070571 0.886024854 0.227194157 0.953475893  0.238962340
#>    25   42 0.215070571 0.886024854 0.110262862 0.984665025  0.158572576
#>    26   38 0.126459136 0.256474787 0.165089772 0.175707343  0.143568955
#>    26   69 0.126459136 0.256474787 0.195210262 0.287009488  0.178582017
#>    26   55 0.126459136 0.256474787 0.098750497 0.213936478  0.077454527
#>    27   51 0.019532606 0.937547590 0.005778683 0.905174578 -0.052660132
#>    27   77 0.019532606 0.937547590 0.057692616 0.907970747  0.031367310
#>    27   57 0.019532606 0.937547590 0.061486149 0.936628334  0.040241463
#>    27   44 0.019532606 0.937547590 0.033344137 1.000000000  0.041132452
#>    27   45 0.019532606 0.937547590 0.007611850 0.990496885  0.030594731
#>    28   96 0.854752695 0.329243854 0.914527284 0.287351136  0.872193236
#>    28   80 0.854752695 0.329243854 0.900826439 0.259193534  0.857108004
#>    30   32 0.573853630 0.770283473 0.574799206 0.827106327  0.655575418
#>    30   90 0.573853630 0.770283473 0.290916500 0.743106880  0.428117223
#>    30   64 0.573853630 0.770283473 0.529637143 0.687820573  0.428703223
#>    33   78 0.389482030 0.262464955 0.488940117 0.229495902  0.432917324
#>    33   89 0.389482030 0.262464955 0.385756391 0.427269552  0.472627792
#>    33   98 0.389482030 0.262464955 0.349568485 0.426105697  0.370306486
#>    33   83 0.389482030 0.262464955 0.275156591 0.364444562  0.357121940
#>    34   56 0.590724609 0.181206863 0.606792692 0.293291532  0.646584609
#>    34   41 0.590724609 0.181206863 0.688585342 0.147539926  0.604305385
#>    34   95 0.590724609 0.181206863 0.716175906 0.203998527  0.652039024
#>    37   75 0.942520528 0.777096737 0.978491797 0.305643237  1.248185329
#>    37   49 0.942520528 0.777096737 1.000000000 0.852088542  1.294480586
#>    37  100 0.942520528 0.777096737 0.918530744 0.973604099  0.897285657
#>    37   99 0.942520528 0.777096737 0.804828527 0.948397027  0.881981562
#>    37   71 0.942520528 0.777096737 0.739585924 0.816128372  0.850133833
#>    37   59 0.942520528 0.777096737 0.793433620 0.716379705  0.843174219
#>    37   54 0.942520528 0.777096737 0.825040919 0.691683649  0.859283849
#>    37   92 0.942520528 0.777096737 0.898345609 0.692717509  0.860649528
#>    38   85 0.165089772 0.175707343 0.072054285 0.136988743  0.114872130
#>    40   42 0.619035173 0.995902369 0.110262862 0.984665025  0.335400604
#>    40   63 0.619035173 0.995902369 0.227194157 0.953475893  0.350098806
#>    40   68 0.619035173 0.995902369 0.563175078 0.969397159  0.372330432
#>    40  100 0.619035173 0.995902369 0.918530744 0.973604099  0.790813807
#>    48   62 0.628619014 0.628089405 0.480116162 0.571428754  0.550355297
#>    48   79 0.628619014 0.628089405 0.619077113 0.527516275  0.560449936
#>    51   52 0.005778683 0.905174578 0.065019084 0.836033201 -0.045935721
#>    53   77 0.071509290 0.906151205 0.057692616 0.907970747  0.066487494
#>    54   81 0.825040919 0.691683649 0.768613429 0.705857969  0.792894642
#>    54   61 0.825040919 0.691683649 0.778967292 0.516481771  0.774836599
#>    72   94 0.303688235 0.480078236 0.331741754 0.424328106  0.297432188
#>    72   98 0.303688235 0.480078236 0.349568485 0.426105697  0.336996694
#>    83   94 0.275156591 0.364444562 0.331741754 0.424328106  0.347919499
#>    40   44 0.619035173 0.995902369 0.033344137 1.000000000  0.335400604
#>          my1           mx2         my2 bp1 bp2
#>   0.66477005  -0.160933910  0.76927661   0   0
#>   0.61960388   0.092227484  0.61237665   0   0
#>   0.60809025   0.204887786  0.50255663   0   0
#>   0.60678661   0.266718462  0.56102090   0   0
#>   0.64822961   0.325837583  0.61297813   0   0
#>   0.66003534   0.249336329  0.81262500   0   0
#>   0.71118799   0.137882927  0.73941226   0   0
#>   0.61316728   0.325837583  0.61297813   0   0
#>   0.50283948   0.365743844  0.48634273   0   0
#>   0.43048180   0.480004777  0.35063776   0   0
#>   0.43635200   0.572416844  0.43165618   0   0
#>   0.48019592   0.560449936  0.58381775   0   0
#>   0.54225179   0.460694295  0.64842284   0   0
#>   0.61732947   0.402669670  0.68261682   0   0
#>   0.59937661   0.702359330  0.57035406   0   0
#>   0.60944793   0.774836599  0.61122707   0   0
#>   0.75110449   0.759300505  0.76236223   0   0
#>   0.77551152   0.655575418  0.79734285   0   0
#>   0.70180513   0.604849106  0.70057789   0   0
#>   0.80901400   0.249336329  0.81262500   0   0
#>   0.84460371   0.153462432  0.89504991   0   0
#>   0.83766679   0.078861267  0.85602842   0   0
#>   0.78548092  -0.045935721  0.80091632   0   0
#>   1.00528973   0.790813807  1.28065697   0   0
#>   0.96407029   0.596645936  0.97097243   0   0
#>   0.87631464   0.555045364  0.89711280   0   0
#>   0.83467549   0.655575418  0.79734285   0   0
#>   0.89557482   0.850133833  0.84382476   0   0
#>   0.21114755   0.658102742  0.24450265   0   0
#>   0.15208464   0.652039024  0.20037049   0   0
#>   0.11290808   0.766879995  0.08920690   0   0
#>   0.17087206   1.025644756  0.13864770   0   0
#>   0.18324543   0.854227941  0.27034038   0   0
#>   0.23696069   0.796528234  0.26134028   0   0
#>   0.23714003   0.723258904  0.30685199   0   0
#>   0.49452993   0.139975397  0.49277651   0   0
#>   0.39033226   0.132821692  0.38010033   0   0
#>   0.38925731   0.213219826  0.33342711   0   0
#>   0.39103948   0.224047430  0.39116284   0   0
#>   0.46582276   0.297432188  0.44199684   0   0
#>   0.52010717   0.266718462  0.56102090   0   0
#>   0.25703933   0.213219826  0.33342711   0   0
#>   0.21328391   0.178582017  0.23178275   0   0
#>   0.08659517   0.146972949  0.08810452   0   0
#>  -0.09313746   0.466308399 -1.91175647   0   1
#>  -0.03737637   0.604305385  0.06162164   0   0
#>   0.04171683   0.557754249  0.24312745   0   0
#>   0.05577453   0.432917324  0.22699401   0   0
#>   0.25470143   0.432917324  0.22699401   0   0
#>   0.26315559   0.327473501  0.30802231   0   0
#>   0.30933594   0.681373525  0.29499943   0   0
#>   0.34919343   0.619740587  0.40689038   0   0
#>   0.30395934   0.529384878  0.29553548   0   0
#>   0.59750937  -0.250239033  0.76608320   0   0
#>   0.59680887 -22.844827600  0.02483583   0   1
#>   0.48002616  -0.094314297  0.42983931   0   0
#>   0.46621596   0.132821692  0.38010033   0   0
#>   0.26796462   0.854227941  0.27034038   0   0
#>   0.34481805   0.804437771  0.33557482   0   0
#>   0.34830080   0.761505578  0.40013368   0   0
#>   0.31789329   0.723258904  0.30685199   0   0
#>  -0.17790172   4.772454782 -1.65931702   0   1
#>   0.13864770   0.960501666  0.24756899   0   0
#>   0.08920690   0.604305385  0.06162164   0   0
#>   0.08723441   2.430384131 -6.55062928   0   1
#>   0.38010033   0.117873341  0.36296449   0   0
#>   0.38261258  -3.276334673  0.11265830   0   1
#>   0.29117177   0.077454527  0.25810186   0   0
#>   0.31303272   0.154076651  0.28695838   0   0
#>   1.32850040   0.372330432  1.44372043   0   0
#>   0.98072456   0.238962340  0.91654563   0   0
#>   0.83170302   0.278195566  0.82794047   0   0
#>   0.80990472   0.428117223  0.80112793   0   0
#>   0.88919084   0.555045364  0.89711280   0   0
#>   0.33557482   0.857108004  0.28061594   0   0
#>   0.34932408   0.884224452  0.30770459   0   0
#>   0.40859763   0.928286763  0.36021929   0   0
#>   0.50974924   1.248185329  0.56331952   0   0
#>   0.51284065   0.863162875  0.58799951   0   0
#>   0.45036532   0.701549100  0.54576022   0   0
#>   0.40013368   0.738818618  0.41156161   0   0
#>   0.76927661  -0.045935721  0.80091632   0   0
#>   0.76608320  -3.717105042  0.83556601   0   1
#>   0.85602842   0.058528453  0.87171402   0   0
#>   0.89504991   0.150809208  0.92274603   0   0
#>   0.98315471   0.335400604  2.31450965   0   0
#>   0.97094628   0.041132452  0.96552416   0   0
#>   0.92442218   0.066487494  0.92138641   0   0
#>   0.91983670   0.060187486  0.87354736   0   0
#>   0.92437423   0.150809208  0.92274603   0   0
#>   0.40689038   0.662403683  0.41491932   0   0
#>   0.43165618   0.700779215  0.54745816   0   0
#>   0.35063776   0.472627792  0.34678899   0   0
#>   0.29553548   0.557754249  0.24312745   0   0
#>   0.17381786  -8.161960323 -3.00403615   0   1
#>   0.17077860   0.114872130  0.16523839   0   0
#>   0.25017566   0.077454527  0.25810186   0   0
#>   0.29499943   0.658102742  0.24450265   0   0
#>   0.41156161   0.704921207  0.51060190   0   0
#>   0.41491932   0.704921207  0.51060190   0   0
#>   0.93099481   0.350098806  1.64904653   0   0
#>   0.21503619   0.114872130  0.16523839   0   0
#>   0.28695838   0.213219826  0.33342711   0   0
#>   0.94911099 -11.767516761  1.20080699   0   1
#>   0.91341129   0.033749121  0.86919037   0   0
#>   0.92486071   0.066487494  0.92138641   0   0
#>   0.96785460 -12.598803973 35.16541174   0   1
#>   0.29053786   0.960501666  0.24756899   0   0
#>   0.30770459   0.928286763  0.36021929   0   0
#>   0.28061594   0.854227941  0.27034038   0   0
#>   0.51060190   0.701549100  0.54576022   0   0
#>   0.79502705   0.402669670  0.68261682   0   0
#>   0.70057789   0.550355297  0.61027496   0   0
#>   0.36021929   0.960501666  0.24756899   0   0
#>   0.89711280   0.596645936  0.97097243   0   0
#>   0.34447588   0.365743844  0.48634273   0   0
#>   0.34126004   0.347919499  0.35236543   0   0
#>   0.30802231   0.224047430  0.39116284   0   0
#>   0.23039303   0.658102742  0.24450265   0   0
#>   0.39116284   0.291190678  0.40596963   0   0
#>   0.56685179   3.121659192  0.49493365   0   1
#>   0.87129245   6.574715131  4.67768535   0   1
#>   0.86942412   0.790813807  1.28065697   0   0
#>   0.80764009   0.759300505  0.76236223   0   0
#>   0.76808389   0.792894642  0.68311553   0   0
#>   0.76620550   0.863162875  0.58799951   0   0
#>   0.08810452  -4.937913256 -3.26792588   0   1
#>   0.61027496   0.460694295  0.64842284   0   0
#>   0.57035406   0.700779215  0.54745816   0   0
#>   0.54576022   0.700779215  0.54745816   0   0
#>   0.86919037   0.058528453  0.87171402   0   0
#>   0.87171402   0.060187486  0.87354736   0   0
#>   0.68311553   0.759300505  0.76236223   0   0
#>   0.64842284   0.402669670  0.68261682   0   0
#>   0.44199684   0.291190678  0.40596963   0   0
#>   0.46190574   0.347919499  0.35236543   0   0
#>   0.66477005   0.097849311  0.61960388   0   0
#>   0.61960388   0.123279848  0.60809025   0   0
#>   0.60809025   0.139569329  0.60678661   0   0
#>   0.60678661   0.278106037  0.64822961   0   0
#>   0.64822961   0.266129276  0.66003534   0   0
#>   0.66003534   0.137545381  0.71118799   0   0
#>   0.71118799   0.100955105  0.66477005   0   0
#>   0.61316728   0.376359062  0.50283948   0   0
#>   0.50283948   0.516899598  0.43048180   0   0
#>   0.43048180   0.524817083  0.43635200   0   0
#>   0.43635200   0.527704827  0.48019592   0   0
#>   0.48019592   0.473156916  0.54225179   0   0
#>   0.54225179   0.335304633  0.61732947   0   0
#>   0.61732947   0.329733221  0.61316728   0   0
#>   0.59937661   0.742295398  0.60944793   0   0
#>   0.60944793   0.716534305  0.75110449   0   0
#>   0.75110449   0.661614406  0.77551152   0   0
#>   0.77551152   0.608035550  0.70180513   0   0
#>   0.70180513   0.723903555  0.59937661   0   0
#>   0.80901400   0.165859250  0.84460371   0   0
#>   0.84460371   0.089705985  0.83766679   0   0
#>   0.83766679   0.069069199  0.78548092   0   0
#>   0.78548092   0.137882927  0.73941226   0   0
#>   0.73941226   0.187656370  0.80901400   0   0
#>   1.00528973   0.655397036  0.96407029   0   0
#>   0.96407029   0.617923634  0.87631464   0   0
#>   0.87631464   0.658062488  0.83467549   0   0
#>   0.83467549   0.745219090  0.89557482   0   0
#>   0.89557482   0.720405396  1.00528973   0   0
#>   0.21114755   0.750846423  0.15208464   0   0
#>   0.15208464   0.776511001  0.11290808   0   0
#>   0.11290808   0.867102487  0.17087206   0   0
#>   0.17087206   0.861798401  0.18324543   0   0
#>   0.18324543   0.784024505  0.23696069   0   0
#>   0.23696069   0.762845965  0.23714003   0   0
#>   0.23714003   0.747740923  0.21114755   0   0
#>   0.50255663   0.197217398  0.49452993   0   0
#>   0.49452993   0.200258527  0.39033226   0   0
#>   0.39033226   0.201694983  0.38925731   0   0
#>   0.38925731   0.218351434  0.39103948   0   0
#>   0.39103948   0.274318642  0.46582276   0   0
#>   0.46582276   0.236379268  0.52010717   0   0
#>   0.52010717   0.204887786  0.50255663   0   0
#>   0.56102090   0.325837583  0.61297813   0   0
#>   0.25703933   0.246939462  0.21328391   0   0
#>   0.21328391   0.248089238  0.08659517   0   0
#>   0.08659517   0.438809218 -0.09313746   0   0
#>  -0.09313746   0.463413426 -0.03737637   0   0
#>  -0.03737637   0.462200152  0.04171683   0   0
#>   0.04171683   0.457590410  0.05577453   0   0
#>   0.05577453   0.309937115  0.25470143   0   0
#>   0.25470143   0.297927795  0.26315559   0   0
#>   0.26315559   0.282730095  0.25703933   0   0
#>   0.30933594   0.645143693  0.34919343   0   0
#>   0.34919343   0.569787526  0.30395934   0   0
#>   0.30395934   0.668530832  0.30933594   0   0
#>   0.59750937   0.071397456  0.59680887   0   0
#>   0.59680887   0.019444151  0.48002616   0   0
#>   0.48002616   0.088444886  0.46621596   0   0
#>   0.46621596   0.139975397  0.49277651   0   0
#>   0.49277651   0.092227484  0.61237665   0   0
#>   0.61237665   0.072008296  0.59750937   0   0
#>   0.26796462   0.771952283  0.34481805   0   0
#>   0.34481805   0.764179450  0.34830080   0   0
#>   0.34830080   0.735930987  0.31789329   0   0
#>   0.31789329   0.796528234  0.26134028   0   0
#>   0.26134028   0.805008373  0.26796462   0   0
#>  -0.17790172   1.025644756  0.13864770   0   0
#>   0.08920690   0.767766608  0.08723441   0   0
#>   0.08723441   1.657666684 -0.17790172   0   0
#>   0.38261258   0.072900501  0.29117177   0   0
#>   0.29117177   0.104969491  0.31303272   0   0
#>   0.31303272   0.117873341  0.36296449   0   0
#>   0.36296449  -0.094314297  0.42983931   0   0
#>   0.42983931  -0.269770989  0.38261258   0   0
#>   1.32850040   0.289561258  0.98072456   0   0
#>   0.98072456   0.326471769  0.83170302   0   0
#>   0.83170302   0.425520216  0.80990472   0   0
#>   0.80990472   0.458072712  0.88919084   0   0
#>   0.88919084   0.377790411  1.32850040   0   0
#>   0.33557482   0.836573472  0.34932408   0   0
#>   0.34932408   0.939763996  0.40859763   0   0
#>   0.40859763   0.989462152  0.50974924   0   0
#>   0.50974924   0.974118343  0.51284065   0   0
#>   0.51284065   0.816747108  0.45036532   0   0
#>   0.45036532   0.761505578  0.40013368   0   0
#>   0.76927661  -0.250239033  0.76608320   0   0
#>   0.89504991   0.078861267  0.85602842   0   0
#>   0.98315471   0.053342267  0.97094628   0   0
#>   0.97094628   0.075718287  0.92442218   0   0
#>   0.92442218   0.123411293  0.91983670   0   0
#>   0.91983670   0.139291611  0.92437423   0   0
#>   0.92437423   0.069973759  0.98315471   0   0
#>   0.40689038   0.572416844  0.43165618   0   0
#>   0.35063776   0.529384878  0.29553548   0   0
#>   0.17381786   0.098903322  0.17077860   0   0
#>   0.17077860   0.074034905  0.25017566   0   0
#>   0.25017566   0.009415969  0.17381786   0   0
#>   0.29499943   0.723258904  0.30685199   0   0
#>   0.41156161   0.662403683  0.41491932   0   0
#>   0.81262500   0.278195566  0.82794047   0   0
#>   0.82794047   0.238962340  0.91654563   0   0
#>   0.91654563   0.158572576  0.93099481   0   0
#>   0.93099481   0.150809208  0.92274603   0   0
#>   0.21503619   0.178582017  0.23178275   0   0
#>   0.23178275   0.154076651  0.28695838   0   0
#>   0.25810186   0.143568955  0.21503619   0   0
#>   0.94911099   0.031367310  0.91341129   0   0
#>   0.91341129   0.040241463  0.92486071   0   0
#>   0.92486071   0.041132452  0.96552416   0   0
#>   0.96552416   0.030594731  0.96785460   0   0
#>   0.96785460  -0.052660132  0.94911099   0   0
#>   0.29053786   0.884224452  0.30770459   0   0
#>   0.28061594   0.872193236  0.29053786   0   0
#>   0.79734285   0.428117223  0.80112793   0   0
#>   0.80112793   0.428703223  0.79502705   0   0
#>   0.79502705   0.604849106  0.70057789   0   0
#>   0.22699401   0.472627792  0.34678899   0   0
#>   0.34678899   0.370306486  0.34447588   0   0
#>   0.34447588   0.357121940  0.34126004   0   0
#>   0.34126004   0.327473501  0.30802231   0   0
#>   0.23039303   0.557754249  0.24312745   0   0
#>   0.06162164   0.652039024  0.20037049   0   0
#>   0.20037049   0.646584609  0.23039303   0   0
#>   0.56331952   1.294480586  0.56685179   0   0
#>   0.56685179   0.897285657  0.87129245   0   0
#>   0.87129245   0.881981562  0.86942412   0   0
#>   0.86942412   0.850133833  0.84382476   0   0
#>   0.84382476   0.843174219  0.80764009   0   0
#>   0.80764009   0.859283849  0.76808389   0   0
#>   0.76808389   0.860649528  0.76620550   0   0
#>   0.76620550   1.248185329  0.56331952   0   0
#>   0.16523839   0.146972949  0.08810452   0   0
#>   2.31450965   0.350098806  1.64904653   0   0
#>   1.64904653   0.372330432  1.44372043   0   0
#>   1.44372043   0.596645936  0.97097243   0   0
#>   1.28065697   1.038037241  4.60119804   0   1
#>   0.61027496   0.560449936  0.58381775   0   0
#>   0.58381775   0.702359330  0.57035406   0   0
#>   0.80091632   0.033749121  0.86919037   0   0
#>   0.92138641   0.060187486  0.87354736   0   0
#>   0.68311553   0.774836599  0.61122707   0   0
#>   0.61122707   0.863162875  0.58799951   0   0
#>   0.44199684   0.336996694  0.46190574   0   0
#>   0.46190574   0.365743844  0.48634273   0   0
#>   0.35236543   0.291190678  0.40596963   0   0
#>   2.31450965   0.347345280  4.02181093   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.183422809 0.634055985
#>   [2,] 0.443328389 0.539091822
#>   [3,] 0.683719461 0.690419492
#>   [4,] 0.106598709 0.819590630
#>   [5,] 0.668680624 0.917605111
#>   [6,] 0.779881603 0.207348143
#>   [7,] 0.250838017 0.443141224
#>   [8,] 0.191193699 0.608079255
#>   [9,] 0.329457673 0.177199079
#>  [10,] 0.605415617 0.318582045
#>  [11,] 0.100965396 0.528803004
#>  [12,] 0.790128630 0.277080977
#>  [13,] 0.863671572 0.076393412
#>  [14,] 0.146685177 0.440101403
#>  [15,] 0.060177991 0.353706129
#>  [16,] 0.331396052 0.927040345
#>  [17,] 0.834602960 0.376339195
#>  [18,] 0.015059471 0.645633270
#>  [19,] 0.103054608 0.858497681
#>  [20,] 0.075067561 0.943160388
#>  [21,] 0.599446304 0.328526417
#>  [22,] 0.074408717 0.206312247
#>  [23,] 0.143448566 0.545763542
#>  [24,] 0.691244639 0.368943855
#>  [25,] 0.215070571 0.886024854
#>  [26,] 0.126459136 0.256474787
#>  [27,] 0.019532606 0.937547590
#>  [28,] 0.854752695 0.329243854
#>  [29,] 0.695166069 0.458187902
#>  [30,] 0.573853630 0.770283473
#>  [31,] 0.858916565 0.334011152
#>  [32,] 0.574799206 0.827106327
#>  [33,] 0.389482030 0.262464955
#>  [34,] 0.590724609 0.181206863
#>  [35,] 0.261125762 0.346990339
#>  [36,] 0.075912801 0.403631229
#>  [37,] 0.942520528 0.777096737
#>  [38,] 0.165089772 0.175707343
#>  [39,] 0.780383685 0.266639149
#>  [40,] 0.619035173 0.995902369
#>  [41,] 0.688585342 0.147539926
#>  [42,] 0.110262862 0.984665025
#>  [43,] 0.808624172 0.285036254
#>  [44,] 0.033344137 1.000000000
#>  [45,] 0.007611850 0.990496885
#>  [46,] 0.163858993 0.326911882
#>  [47,] 0.712209092 0.008312603
#>  [48,] 0.628619014 0.628089405
#>  [49,] 1.000000000 0.852088542
#>  [50,] 0.744746717 0.475157316
#>  [51,] 0.005778683 0.905174578
#>  [52,] 0.065019084 0.836033201
#>  [53,] 0.071509290 0.906151205
#>  [54,] 0.825040919 0.691683649
#>  [55,] 0.098750497 0.213936478
#>  [56,] 0.606792692 0.293291532
#>  [57,] 0.061486149 0.936628334
#>  [58,] 0.001088476 0.573235405
#>  [59,] 0.793433620 0.716379705
#>  [60,] 0.612647717 0.941532274
#>  [61,] 0.778967292 0.516481771
#>  [62,] 0.480116162 0.571428754
#>  [63,] 0.227194157 0.953475893
#>  [64,] 0.529637143 0.687820573
#>  [65,] 0.162467003 0.000000000
#>  [66,] 0.094688224 0.874492507
#>  [67,] 0.459630210 0.569024062
#>  [68,] 0.563175078 0.969397159
#>  [69,] 0.195210262 0.287009488
#>  [70,] 0.030713736 0.243289779
#>  [71,] 0.739585924 0.816128372
#>  [72,] 0.303688235 0.480078236
#>  [73,] 0.051832779 0.737785622
#>  [74,] 0.378467486 0.213576565
#>  [75,] 0.978491797 0.305643237
#>  [76,] 0.728527356 0.237191695
#>  [77,] 0.057692616 0.907970747
#>  [78,] 0.488940117 0.229495902
#>  [79,] 0.619077113 0.527516275
#>  [80,] 0.900826439 0.259193534
#>  [81,] 0.768613429 0.705857969
#>  [82,] 0.672558887 0.468233633
#>  [83,] 0.275156591 0.364444562
#>  [84,] 0.810251584 0.251320595
#>  [85,] 0.072054285 0.136988743
#>  [86,] 0.223869562 0.735728165
#>  [87,] 0.178527722 0.572889854
#>  [88,] 0.853966452 0.043819248
#>  [89,] 0.385756391 0.427269552
#>  [90,] 0.290916500 0.743106880
#>  [91,] 0.000000000 0.616845442
#>  [92,] 0.898345609 0.692717509
#>  [93,] 0.259178490 0.436899280
#>  [94,] 0.331741754 0.424328106
#>  [95,] 0.716175906 0.203998527
#>  [96,] 0.914527284 0.287351136
#>  [97,] 0.277619210 0.913720588
#>  [98,] 0.349568485 0.426105697
#>  [99,] 0.804828527 0.948397027
#> [100,] 0.918530744 0.973604099
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.1834228,0.634056) [7]: 8 18 23 73 86 87 90 
#> 2: (0.4433284,0.5390918) [7]: 21 62 67 72 79 89 90 
#> 3: (0.6837195,0.6904195) [5]: 30 48 61 71 81 
#> 4: (0.1065987,0.8195906) [5]: 19 25 52 73 86 
#> 5: (0.6686806,0.9176051) [5]: 32 40 60 71 99 
#> 6: (0.7798816,0.2073481) [7]: 13 39 41 76 80 84 95 
#> 7: (0.250838,0.4431412) [7]: 14 23 35 46 72 87 93 
#> 8: (0.1911937,0.6080793) [4]: 1 72 87 90 
#> 9: (0.3294577,0.1771991) [9]: 33 34 35 38 47 65 69 74 78 
#> 10: (0.6054156,0.318582) [3]: 21 24 56 
#> 11: (0.1009654,0.528803) [6]: 14 18 23 36 58 91 
#> 12: (0.7901286,0.277081) [5]: 17 24 39 43 84 
#> 13: (0.8636716,0.07639341) [6]: 6 41 47 75 80 88 
#> 14: (0.1466852,0.4401014) [5]: 7 11 23 36 46 
#> 15: (0.06017799,0.3537061) [5]: 26 36 46 58 70 
#> 16: (0.3313961,0.9270403) [5]: 32 63 68 90 97 
#> 17: (0.834603,0.3763392) [9]: 12 24 28 31 43 50 61 75 92 
#> 18: (0.01505947,0.6456333) [6]: 1 11 23 51 73 91 
#> 19: (0.1030546,0.8584977) [4]: 4 25 52 66 
#> 20: (0.07506756,0.9431604) [5]: 42 44 53 57 66 
#> 21: (0.5994463,0.3285264) [8]: 2 10 24 56 78 79 82 89 
#> 22: (0.07440872,0.2063122) [3]: 55 70 85 
#> 23: (0.1434486,0.5457635) [6]: 1 7 11 14 18 87 
#> 24: (0.6912446,0.3689439) [10]: 10 12 17 21 29 39 50 56 76 82 
#> 25: (0.2150706,0.8860249) [8]: 4 19 42 63 66 86 90 97 
#> 26: (0.1264591,0.2564748) [6]: 15 38 46 55 69 70 
#> 27: (0.01953261,0.9375476) [5]: 44 45 51 57 77 
#> 28: (0.8547527,0.3292439) [5]: 17 31 43 80 96 
#> 29: (0.6951661,0.4581879) [3]: 24 50 82 
#> 30: (0.5738536,0.7702835) [6]: 3 32 48 64 71 90 
#> 31: (0.8589166,0.3340112) [4]: 17 28 75 96 
#> 32: (0.5747992,0.8271063) [7]: 5 16 30 60 68 71 90 
#> 33: (0.389482,0.262465) [7]: 9 35 74 78 83 89 98 
#> 34: (0.5907246,0.1812069) [6]: 9 41 47 56 78 95 
#> 35: (0.2611258,0.3469903) [7]: 7 9 33 46 69 83 93 
#> 36: (0.0759128,0.4036312) [5]: 11 14 15 46 58 
#> 37: (0.9425205,0.7770967) [8]: 49 54 59 71 75 92 99 100 
#> 38: (0.1650898,0.1757073) [6]: 9 26 55 65 69 85 
#> 39: (0.7803837,0.2666391) [5]: 6 12 24 76 84 
#> 40: (0.6190352,0.9959024) [8]: 5 42 44 60 63 68 99 100 
#> 41: (0.6885853,0.1475399) [5]: 6 13 34 47 95 
#> 42: (0.1102629,0.984665) [6]: 20 25 40 44 63 66 
#> 43: (0.8086242,0.2850363) [5]: 12 17 28 80 84 
#> 44: (0.03334414,1) [6]: 20 27 40 42 45 57 
#> 45: (0.00761185,0.9904969) [3]: 27 44 51 
#> 46: (0.163859,0.3269119) [7]: 7 14 15 26 35 36 69 
#> 47: (0.7122091,0.008312603) [6]: 9 13 34 41 65 88 
#> 48: (0.628619,0.6280894) [6]: 3 30 61 62 64 79 
#> 49: (1,0.8520885) [3]: 37 75 100 
#> 50: (0.7447467,0.4751573) [5]: 17 24 29 61 82 
#> 51: (0.005778683,0.9051746) [7]: 18 27 45 52 73 77 91 
#> 52: (0.06501908,0.8360332) [6]: 4 19 51 66 73 77 
#> 53: (0.07150929,0.9061512) [4]: 20 57 66 77 
#> 54: (0.8250409,0.6916836) [5]: 37 59 61 81 92 
#> 55: (0.0987505,0.2139365) [5]: 22 26 38 70 85 
#> 56: (0.6067927,0.2932915) [7]: 10 21 24 34 76 78 95 
#> 57: (0.06148615,0.9366283) [5]: 20 27 44 53 77 
#> 58: (0.001088476,0.5732354) [5]: 11 15 36 70 91 
#> 59: (0.7934336,0.7163797) [4]: 37 54 71 81 
#> 60: (0.6126477,0.9415323) [4]: 5 32 40 68 
#> 61: (0.7789673,0.5164818) [9]: 3 17 48 50 54 79 81 82 92 
#> 62: (0.4801162,0.5714288) [5]: 2 48 64 67 79 
#> 63: (0.2271942,0.9534759) [6]: 16 25 40 42 68 97 
#> 64: (0.5296371,0.6878206) [5]: 30 48 62 67 90 
#> 65: (0.162467,0) [4]: 9 38 47 85 
#> 66: (0.09468822,0.8744925) [7]: 19 20 25 42 52 53 77 
#> 67: (0.4596302,0.5690241) [4]: 2 62 64 90 
#> 68: (0.5631751,0.9693972) [5]: 16 32 40 60 63 
#> 69: (0.1952103,0.2870095) [5]: 9 26 35 38 46 
#> 70: (0.03071374,0.2432898) [6]: 15 22 26 55 58 85 
#> 71: (0.7395859,0.8161284) [8]: 3 5 30 32 37 59 81 99 
#> 72: (0.3036882,0.4800782) [9]: 2 7 8 87 89 90 93 94 98 
#> 73: (0.05183278,0.7377856) [6]: 1 4 18 51 52 86 
#> 74: (0.3784675,0.2135766) [3]: 9 33 78 
#> 75: (0.9784918,0.3056432) [9]: 13 17 31 37 49 80 88 92 96 
#> 76: (0.7285274,0.2371917) [5]: 6 24 39 56 95 
#> 77: (0.05769262,0.9079707) [6]: 27 51 52 53 57 66 
#> 78: (0.4889401,0.2294959) [7]: 9 21 33 34 56 74 89 
#> 79: (0.6190771,0.5275163) [6]: 2 21 48 61 62 82 
#> 80: (0.9008264,0.2591935) [7]: 6 13 28 43 75 84 96 
#> 81: (0.7686134,0.705858) [5]: 3 54 59 61 71 
#> 82: (0.6725589,0.4682336) [6]: 21 24 29 50 61 79 
#> 83: (0.2751566,0.3644446) [5]: 33 35 93 94 98 
#> 84: (0.8102516,0.2513206) [5]: 6 12 39 43 80 
#> 85: (0.07205428,0.1369887) [5]: 22 38 55 65 70 
#> 86: (0.2238696,0.7357282) [5]: 1 4 25 73 90 
#> 87: (0.1785277,0.5728899) [5]: 1 7 8 23 72 
#> 88: (0.8539665,0.04381925) [3]: 13 47 75 
#> 89: (0.3857564,0.4272696) [6]: 2 21 33 72 78 98 
#> 90: (0.2909165,0.7431069) [12]: 1 2 8 16 25 30 32 64 67 72 86 97 
#> 91: (0,0.6168454) [4]: 11 18 51 58 
#> 92: (0.8983456,0.6927175) [5]: 17 37 54 61 75 
#> 93: (0.2591785,0.4368993) [5]: 7 35 72 83 94 
#> 94: (0.3317418,0.4243281) [4]: 72 83 93 98 
#> 95: (0.7161759,0.2039985) [5]: 6 34 41 56 76 
#> 96: (0.9145273,0.2873511) [4]: 28 31 75 80 
#> 97: (0.2776192,0.9137206) [4]: 16 25 63 90 
#> 98: (0.3495685,0.4261057) [5]: 33 72 83 89 94 
#> 99: (0.8048285,0.948397) [5]: 5 37 40 71 100 
#> 100: (0.9185307,0.9736041) [4]: 37 40 49 99 
#> number of nodes: 100 
#> number of arcs: 283 
#> number of boundary nodes: 14 
#> boundary nodes:  40 44 45 47 49 51 58 65 70 75 85 88 91 100 
#> number of triangles: 184 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.099218608 0.162658298 0.044343928 0.037399477 0.167860348 0.067451740 
#>                                                                         
#> 0.172049086 0.263338536 0.097590672 0.235400904 0.199955826 0.145734182 
#>                                                                         
#> 0.020626604 0.242423480 0.187245478 0.189659028 0.114027005 0.171956238 
#>                                                                         
#> 0.152375852 0.150554034 0.115348695 0.044174132 0.099128529 0.191770508 
#>                                                                         
#> 0.054744012 0.120523056 0.165151984 0.147484218 0.035416729 0.062839580 
#>                                                                         
#> 0.188989456 0.186537848 0.090916376 0.033567090 0.059634147 0.105711699 
#>                                                                         
#> 0.114484967 0.099317515 0.089930026 0.062012415 0.155817683 0.089121017 
#>                                                                         
#> 0.115305730 0.175726917 0.549804932 0.211307607 0.112658395 0.111613743 
#>                                                                         
#> 0.050113820 0.153687579 0.113381680 0.100302062 0.035992590 0.032488871 
#>                                                                         
#> 0.043623619 0.185375986 0.079616599 0.033754913 0.094926943 0.143548944 
#>                                                                         
#> 0.135690907 0.289928204 0.090495761 0.141217312 0.146136480 0.116706430 
#>                                                                         
#> 0.331272956 0.096648983 0.079750447 0.336357943 0.064211916 0.171131106 
#>                                                                         
#> 0.296049486 0.142764845 0.063891726 0.006329687 0.122894161 0.395284585 
#>                                                                         
#> 0.212861969 0.053654062 0.118927590 0.173608876 0.288387038 0.048573409 
#>                                                                         
#> 0.120933471 0.078432466 0.069339317 0.032082997 0.039236923 0.111267934 
#>                                                                         
#> 0.101032754 0.079841917 0.223072406 0.134011611 0.114056790 0.081447171 
#>                                                                         
#> 0.074098699 0.134039249 0.052404215 0.024738661 0.121019377 0.076566090 
#>                                                                         
#> 0.050745473 0.085341998 0.051989182 0.028907580 0.027431001 0.031313954 
#>                                                                         
#> 0.072592762 0.095755440 0.072519102 0.245039020 0.115608059 0.066528640 
#>                                                                         
#> 0.056780223 0.036206617 0.096639669 0.022394510 0.141201729 0.074195592 
#>                                                                         
#> 0.546868425 0.146298565 0.116462830 0.113355066 0.040111301 0.073311981 
#>                                                                         
#> 0.164135235 0.126488668 0.160270489 0.116835917 0.072309665 0.049894564 
#>                                                                         
#> 0.026958279 0.137889745 0.073644155 0.017915138 0.168760916 0.096920047 
#>                                                                         
#> 0.061361694 0.027114152 0.153124104 0.109421990 0.167558270 0.151598073 
#>                                                                         
#> 0.125772632 0.262127056 0.176129519 0.048979765 0.034083549 0.254659631 
#>                                                                         
#> 0.198309052 0.086286339 0.137563745 0.135826199 0.083193143 0.127199257 
#>                                                                         
#> 0.039068137 0.044712675 0.098444737 0.144171308 0.092709932 0.060927792 
#>                                                                         
#> 0.130398429 0.123794548 0.139586510 0.063793696 0.109142246 0.155466718 
#>                                                                         
#> 0.131588746 0.053440736 0.059293133 0.059396096 0.148539000 0.104197191 
#>                                                                         
#> 0.145170969 0.096699692 0.010417551 0.064478589 0.148537826 0.170409159 
#>                                                                         
#> 0.173437866 0.164374670 0.243485928 0.418355460 0.261297674 0.167838040 
#>                                                                         
#> 0.061035098 0.104274604 0.183025468 0.099513481 0.011598414 0.025327977 
#>                                                                         
#> 0.133960748 0.109314397 0.127654243 0.099791144 0.045743628 0.145014271 
#>                                                                         
#> 0.020133840 0.108766538 0.134969744 0.014282706 0.032688386 0.256396502 
#>                                                                         
#> 0.166059867 0.033989198 0.114279975 0.117673804 0.107087263 0.052345963 
#>                                                                         
#> 0.227342635 0.107502898 0.055401848 0.188335110 0.263119560 0.235617522 
#>                                                                         
#> 0.051224827 0.048814082 0.160318171 0.322735748 0.150782191 0.133563339 
#>                                                                         
#> 0.259707188 0.018050784 0.070509472 0.015070583 0.037179845 0.071416023 
#>                                                                         
#> 0.054418233 0.157681819 0.148386859 0.069363474 0.025507865 0.057241499 
#>                                                                         
#> 0.136925647 0.089330161 0.161796639 0.068406037 0.068531919 0.143925464 
#>                                                                         
#> 0.089530475 0.075226892 0.050766883 0.035173603 0.048280183 0.041963613 
#>                                                                         
#> 0.063961409 0.054274600 0.072993160 0.083844125 0.056830721 0.284239313 
#>                                                                         
#> 0.093569373 0.104780101 0.164846704 0.168438070 0.153199694 0.113230545 
#>                                                                         
#> 0.103490027 0.127504854 0.472823788 0.094486298 0.197966293 0.219779154 
#>                                                                         
#> 0.206654112 0.160976594 0.145247562 0.095243254 0.100770690 0.508896396 
#>                                                                         
#> 0.394131181 0.061829414 0.300324508 0.158945042 0.101024762 0.091049190 
#>                                                                         
#> 0.013935968 0.058180520 0.181158707 0.062410552 0.070838071 0.082388831 
#>             
#> 0.585705370 
#> 
#> $alpha
#> [1] 0.2985572
#> 
#> attr(,"class")
#> [1] "scree"