Pre-processing to generate scagnostic measures

Pre-processing to generate scagnostic measures

scree(x, y, binner = NULL, ...)

Arguments

x, y

numeric vectors

binner

an optional function that bins the x and y vectors prior to triangulation

...

other args

Value

An object of class "scree" that consists of three elements:

del: the Delauney-Vornoi tesselation from alphahull::delvor()

weights: the lengths of each edge in the Delauney triangulation

alpha: the radius or alpha value that will be used to generate the alphahull

Examples

x <- runif(100)
y <- runif(100)
scree(x,y)
#> $del
#> $mesh
#>  ind1 ind2          x1          y1          x2          y2          mx1
#>    27   76 0.770443166 0.710702061 0.814197680 0.759739423   0.77316085
#>    76   10 0.814197680 0.759739423 0.775901911 0.838179689   0.77478029
#>    10   13 0.775901911 0.838179689 0.644739798 0.927696709   0.68242837
#>    13   36 0.644739798 0.927696709 0.623396784 0.735029583   0.65761316
#>    36   27 0.623396784 0.735029583 0.770443166 0.710702061   0.69225069
#>    45   61 0.322225162 0.237624211 0.323579525 0.127254642   0.37211028
#>    61   86 0.323579525 0.127254642 0.406186840 0.094917281   0.38793454
#>    86   22 0.406186840 0.094917281 0.535002927 0.188181623   0.45377838
#>    22   56 0.535002927 0.188181623 0.453304542 0.267185767   0.47176230
#>    56   96 0.453304542 0.267185767 0.402920826 0.295191783   0.42280407
#>    96   45 0.402920826 0.295191783 0.322225162 0.237624211   0.36931606
#>    75   55 0.338789736 0.690845038 0.339691903 0.749124662   0.29797287
#>    55   43 0.339691903 0.749124662 0.386064477 0.837061138   0.33480500
#>    43   67 0.386064477 0.837061138 0.380960267 0.871113149   0.31052346
#>    67   77 0.380960267 0.871113149 0.218619241 0.872223266   0.29966232
#>    77   58 0.218619241 0.872223266 0.164330003 0.725060761   0.21094283
#>    58    6 0.164330003 0.725060761 0.241547212 0.686964806   0.22541509
#>     6   75 0.241547212 0.686964806 0.338789736 0.690845038   0.28912256
#>    99   42 0.220414505 0.945394707 0.003508729 0.868628470  -0.05322153
#>    42   53 0.003508729 0.868628470 0.156045936 0.691772411   0.02898069
#>    53   58 0.156045936 0.691772411 0.164330003 0.725060761   0.03362086
#>    77   99 0.218619241 0.872223266 0.220414505 0.945394707   0.11905034
#>    67   81 0.380960267 0.871113149 0.388706036 0.985323385   0.33609662
#>    81   99 0.388706036 0.985323385 0.220414505 0.945394707   0.17782399
#>    58   29 0.164330003 0.725060761 0.166023769 0.676489733   0.20097042
#>    29   66 0.166023769 0.676489733 0.257183710 0.611847019   0.20942169
#>    66   75 0.257183710 0.611847019 0.338789736 0.690845038   0.29139548
#>    97   34 0.678276151 0.177059113 0.718636695 0.160760272   0.71852881
#>    34   90 0.718636695 0.160760272 0.769346416 0.184345754   0.72400999
#>    90   98 0.769346416 0.184345754 0.913519875 0.236911924   0.83888158
#>    98   80 0.913519875 0.236911924 0.803430663 0.381820930   0.84530128
#>    80   72 0.803430663 0.381820930 0.751891435 0.397020281   0.75790855
#>    72   20 0.751891435 0.397020281 0.695003463 0.364346427   0.75695826
#>    20   97 0.695003463 0.364346427 0.678276151 0.177059113   0.67651516
#>    41   18 0.248035250 0.593959346 0.212554969 0.406119080   0.28545851
#>    18   17 0.212554969 0.406119080 0.271721610 0.270166247   0.22066804
#>    17  100 0.271721610 0.270166247 0.374982114 0.353656427   0.31562492
#>   100   48 0.374982114 0.353656427 0.356671577 0.394844640   0.32481938
#>    48   32 0.356671577 0.394844640 0.401579185 0.502992443   0.29849881
#>    32   41 0.401579185 0.502992443 0.248035250 0.593959346   0.29037217
#>    85   81 0.757352368 1.000000000 0.388706036 0.985323385   0.53084772
#>    67   39 0.380960267 0.871113149 0.420880882 0.844065074   0.44185938
#>    39   83 0.420880882 0.844065074 0.514267130 0.878910172   0.44763278
#>    83   13 0.514267130 0.878910172 0.644739798 0.927696709   0.56497596
#>    13   85 0.644739798 0.927696709 0.757352368 1.000000000   0.61480766
#>    76   78 0.814197680 0.759739423 0.917746909 0.796350520   0.84922607
#>    78   54 0.917746909 0.796350520 0.967975720 0.946602668   0.86999439
#>    54   24 0.967975720 0.946602668 0.752007560 0.929325163   0.86226672
#>    24   13 0.752007560 0.929325163 0.644739798 0.927696709   0.69931044
#>    50   82 0.999658163 0.674645467 0.866343397 0.597037594   0.96271591
#>    82   57 0.866343397 0.597037594 0.872305019 0.521505965   0.94380519
#>    57   84 0.872305019 0.521505965 0.966134396 0.469611516   0.92657240
#>    84   37 0.966134396 0.469611516 0.982088838 0.481435555   0.97035411
#>    37   50 0.982088838 0.481435555 0.999658163 0.674645467   1.15049177
#>    98   89 0.913519875 0.236911924 0.969692975 0.219022641   0.93894418
#>    89   84 0.969692975 0.219022641 0.966134396 0.469611516   1.05581929
#>    84   93 0.966134396 0.469611516 0.907670486 0.416156342   1.02159419
#>    93   80 0.907670486 0.416156342 0.803430663 0.381820930   0.87960540
#>    24   85 0.752007560 0.929325163 0.757352368 1.000000000   0.69775950
#>    83   26 0.514267130 0.878910172 0.522718789 0.852089056   0.58571817
#>    26   95 0.522718789 0.852089056 0.559950954 0.775697329   0.61006464
#>    95   36 0.559950954 0.775697329 0.623396784 0.735029583   0.63996942
#>    49   47 0.230178662 0.247645090 0.137534976 0.384636663   0.10923946
#>    47   31 0.137534976 0.384636663 0.094788192 0.109384206   0.08154551
#>    31   87 0.094788192 0.109384206 0.245446180 0.019742202   0.19834154
#>    87   61 0.245446180 0.019742202 0.323579525 0.127254642   0.22641007
#>    45   49 0.322225162 0.237624211 0.230178662 0.247645090   0.27328135
#>    89   65 0.969692975 0.219022641 0.799389037 0.125840015   0.91436611
#>    65   79 0.799389037 0.125840015 0.756920067 0.066998588   0.84042613
#>    79   33 0.756920067 0.066998588 0.412804968 0.000000000   0.66346211
#>    35   91 0.588700301 0.566453094 0.509341582 0.468798987   0.57567863
#>    91   60 0.509341582 0.468798987 0.537330249 0.386097673   0.58633827
#>    60   19 0.537330249 0.386097673 0.561675206 0.365624646   0.59984758
#>    19   59 0.561675206 0.365624646 0.666038543 0.376036057   0.60789407
#>    59   64 0.666038543 0.376036057 0.728078340 0.468888663   0.67063906
#>    64   27 0.728078340 0.468888663 0.770443166 0.710702061   0.72949006
#>    27   35 0.770443166 0.710702061 0.588700301 0.566453094   0.70346821
#>    45  100 0.322225162 0.237624211 0.374982114 0.353656427   0.32947579
#>    18   49 0.212554969 0.406119080 0.230178662 0.247645090   0.21451219
#>    41   47 0.248035250 0.593959346 0.137534976 0.384636663   0.14020156
#>    60   56 0.537330249 0.386097673 0.453304542 0.267185767   0.50334253
#>    22   97 0.535002927 0.188181623 0.678276151 0.177059113   0.61320195
#>    97   59 0.678276151 0.177059113 0.666038543 0.376036057   0.62356080
#>    72   64 0.751891435 0.397020281 0.728078340 0.468888663   0.70086932
#>    80   94 0.803430663 0.381820930 0.781783347 0.446608092   0.78985736
#>    94   64 0.781783347 0.446608092 0.728078340 0.468888663   0.75638278
#>    86   68 0.406186840 0.094917281 0.554109783 0.143858825   0.47468523
#>    68   97 0.554109783 0.143858825 0.678276151 0.177059113   0.60745582
#>    66   41 0.257183710 0.611847019 0.248035250 0.593959346   0.29414267
#>    32   71 0.401579185 0.502992443 0.466256784 0.654360851   0.42197118
#>    71   75 0.466256784 0.654360851 0.338789736 0.690845038   0.39433651
#>    54   51 0.967975720 0.946602668 0.959902603 0.964856359   0.86217114
#>    51   85 0.959902603 0.964856359 0.757352368 1.000000000   0.85423870
#>    27   82 0.770443166 0.710702061 0.866343397 0.597037594   0.76896319
#>    82   38 0.866343397 0.597037594 0.861718974 0.650491997   0.83581912
#>    38   52 0.861718974 0.650491997 0.867898308 0.710695995   0.88197010
#>    52   27 0.867898308 0.710695995 0.770443166 0.710702061   0.81916938
#>    39   95 0.420880882 0.844065074 0.559950954 0.775697329   0.47696516
#>    36   69 0.623396784 0.735029583 0.609199272 0.713156247   0.68928530
#>    69   35 0.609199272 0.713156247 0.588700301 0.566453094   0.68853294
#>    64   82 0.728078340 0.468888663 0.866343397 0.597037594   0.74362097
#>    52   76 0.867898308 0.710695995 0.814197680 0.759739423   0.81917077
#>    97   73 0.678276151 0.177059113 0.590156032 0.099604268   0.66059815
#>    73   79 0.590156032 0.099604268 0.756920067 0.066998588   0.67230869
#>    65   90 0.799389037 0.125840015 0.769346416 0.184345754   0.76116175
#>    53   30 0.156045936 0.691772411 0.007691770 0.571691024   0.08149124
#>    30   41 0.007691770 0.571691024 0.248035250 0.593959346   0.13009218
#>    70   46 0.000000000 0.053160216 0.051185490 0.405536474  -0.60891096
#>    46   47 0.051185490 0.405536474 0.137534976 0.384636663   0.12292005
#>    42   70 0.003508729 0.868628470 0.000000000 0.053160216 -14.09347368
#>    70   92 0.000000000 0.053160216 0.094388846 0.102442214   0.03059460
#>    92   62 0.094388846 0.102442214 0.215666137 0.009678677   0.18900486
#>    62   87 0.215666137 0.009678677 0.245446180 0.019742202   0.19791939
#>    48   91 0.356671577 0.394844640 0.509341582 0.468798987   0.43620647
#>    91   40 0.509341582 0.468798987 0.499491458 0.515347023   0.45403776
#>    40   71 0.499491458 0.515347023 0.466256784 0.654360851   0.44221211
#>    73   86 0.590156032 0.099604268 0.406186840 0.094917281   0.49928075
#>    69   88 0.609199272 0.713156247 0.505748898 0.672198752   0.57720420
#>    88   71 0.505748898 0.672198752 0.466256784 0.654360851   0.51667895
#>    95   69 0.559950954 0.775697329 0.609199272 0.713156247   0.58475387
#>    89   50 0.969692975 0.219022641 0.999658163 0.674645467   6.22063936
#>    50   52 0.999658163 0.674645467 0.867898308 0.710695995   0.92868023
#>    55   71 0.339691903 0.749124662 0.466256784 0.654360851   0.43741492
#>    88   95 0.505748898 0.672198752 0.559950954 0.775697329   0.46550803
#>    78   50 0.917746909 0.796350520 0.999658163 0.674645467   1.00739851
#>    96  100 0.402920826 0.295191783 0.374982114 0.353656427   0.33841369
#>   100   91 0.374982114 0.353656427 0.509341582 0.468798987   0.44404950
#>    78   52 0.917746909 0.796350520 0.867898308 0.710695995   0.94251528
#>    60   96 0.537330249 0.386097673 0.402920826 0.295191783   0.46520530
#>    94   63 0.781783347 0.446608092 0.877625580 0.432315711   0.83493288
#>    63   84 0.877625580 0.432315711 0.966134396 0.469611516   0.91006408
#>    92   74 0.094388846 0.102442214 0.107024285 0.018593816   0.16595819
#>     1   76 0.732936104 0.763320582 0.814197680 0.759739423   0.77316085
#>     1   10 0.732936104 0.763320582 0.775901911 0.838179689   0.77478029
#>     1   13 0.732936104 0.763320582 0.644739798 0.927696709   0.68242837
#>     1   36 0.732936104 0.763320582 0.623396784 0.735029583   0.65761316
#>     1   27 0.732936104 0.763320582 0.770443166 0.710702061   0.69225069
#>     2   61 0.417732104 0.241234791 0.323579525 0.127254642   0.37211028
#>     2   86 0.417732104 0.241234791 0.406186840 0.094917281   0.38793454
#>     2   22 0.417732104 0.241234791 0.535002927 0.188181623   0.45377838
#>     2   56 0.417732104 0.241234791 0.453304542 0.267185767   0.47176230
#>     2   96 0.417732104 0.241234791 0.402920826 0.295191783   0.42280407
#>     2   45 0.417732104 0.241234791 0.322225162 0.237624211   0.36931606
#>     3   55 0.289287958 0.770396726 0.339691903 0.749124662   0.29797287
#>     3   43 0.289287958 0.770396726 0.386064477 0.837061138   0.33480500
#>     3   67 0.289287958 0.770396726 0.380960267 0.871113149   0.31052346
#>     3   77 0.289287958 0.770396726 0.218619241 0.872223266   0.29966232
#>     3   58 0.289287958 0.770396726 0.164330003 0.725060761   0.21094283
#>     3    6 0.289287958 0.770396726 0.241547212 0.686964806   0.22541509
#>     3   75 0.289287958 0.770396726 0.338789736 0.690845038   0.28912256
#>     4   42 0.020066181 0.870763954 0.003508729 0.868628470  -0.05322153
#>     4   53 0.020066181 0.870763954 0.156045936 0.691772411   0.02898069
#>     4   58 0.020066181 0.870763954 0.164330003 0.725060761   0.03362086
#>     4   77 0.020066181 0.870763954 0.218619241 0.872223266   0.11968350
#>     4   99 0.020066181 0.870763954 0.220414505 0.945394707   0.11905034
#>     5   77 0.263246131 0.950364352 0.218619241 0.872223266   0.24644022
#>     5   67 0.263246131 0.950364352 0.380960267 0.871113149   0.29983070
#>     5   81 0.263246131 0.950364352 0.388706036 0.985323385   0.33609662
#>     5   99 0.263246131 0.950364352 0.220414505 0.945394707   0.17782399
#>     6   29 0.241547212 0.686964806 0.166023769 0.676489733   0.20097042
#>     6   66 0.241547212 0.686964806 0.257183710 0.611847019   0.20942169
#>     7   34 0.768169570 0.248328717 0.718636695 0.160760272   0.71852881
#>     7   90 0.768169570 0.248328717 0.769346416 0.184345754   0.72400999
#>     7   98 0.768169570 0.248328717 0.913519875 0.236911924   0.83888158
#>     7   80 0.768169570 0.248328717 0.803430663 0.381820930   0.84530128
#>     7   72 0.768169570 0.248328717 0.751891435 0.397020281   0.75790855
#>     7   20 0.768169570 0.248328717 0.695003463 0.364346427   0.75695826
#>     7   97 0.768169570 0.248328717 0.678276151 0.177059113   0.67651516
#>     8   18 0.280116055 0.378900504 0.212554969 0.406119080   0.28545851
#>     8   17 0.280116055 0.378900504 0.271721610 0.270166247   0.22066804
#>     8  100 0.280116055 0.378900504 0.374982114 0.353656427   0.31562492
#>     8   48 0.280116055 0.378900504 0.356671577 0.394844640   0.32481938
#>     8   32 0.280116055 0.378900504 0.401579185 0.502992443   0.29849881
#>     8   41 0.280116055 0.378900504 0.248035250 0.593959346   0.29037217
#>     9   81 0.491599796 0.976611775 0.388706036 0.985323385   0.53084772
#>     9   67 0.491599796 0.976611775 0.380960267 0.871113149   0.43539650
#>     9   39 0.491599796 0.976611775 0.420880882 0.844065074   0.44185938
#>     9   83 0.491599796 0.976611775 0.514267130 0.878910172   0.44763278
#>     9   13 0.491599796 0.976611775 0.644739798 0.927696709   0.56497596
#>     9   85 0.491599796 0.976611775 0.757352368 1.000000000   0.61480766
#>    10   78 0.775901911 0.838179689 0.917746909 0.796350520   0.84922607
#>    10   54 0.775901911 0.838179689 0.967975720 0.946602668   0.86999439
#>    10   24 0.775901911 0.838179689 0.752007560 0.929325163   0.86226672
#>    11   82 0.978624040 0.488961469 0.866343397 0.597037594   0.96271591
#>    11   57 0.978624040 0.488961469 0.872305019 0.521505965   0.94380519
#>    11   84 0.978624040 0.488961469 0.966134396 0.469611516   0.92657240
#>    11   37 0.978624040 0.488961469 0.982088838 0.481435555   0.97035411
#>    11   50 0.978624040 0.488961469 0.999658163 0.674645467   1.15049177
#>    12   89 0.936063778 0.250226896 0.969692975 0.219022641   0.93894418
#>    12   84 0.936063778 0.250226896 0.966134396 0.469611516   1.05581929
#>    12   93 0.936063778 0.250226896 0.907670486 0.416156342   1.02159419
#>    12   80 0.936063778 0.250226896 0.803430663 0.381820930   0.87960540
#>    12   98 0.936063778 0.250226896 0.913519875 0.236911924   0.87742686
#>    13   26 0.644739798 0.927696709 0.522718789 0.852089056   0.58571817
#>    13   95 0.644739798 0.927696709 0.559950954 0.775697329   0.61006464
#>    14   47 0.219751917 0.213348346 0.137534976 0.384636663   0.10923946
#>    14   31 0.219751917 0.213348346 0.094788192 0.109384206   0.08154551
#>    14   87 0.219751917 0.213348346 0.245446180 0.019742202   0.19834154
#>    14   61 0.219751917 0.213348346 0.323579525 0.127254642   0.22641007
#>    14   45 0.219751917 0.213348346 0.322225162 0.237624211   0.28130724
#>    14   49 0.219751917 0.213348346 0.230178662 0.247645090   0.27328135
#>    15   65 0.910001080 0.002755188 0.799389037 0.125840015   0.91436611
#>    15   79 0.910001080 0.002755188 0.756920067 0.066998588   0.84042613
#>    15   33 0.910001080 0.002755188 0.412804968 0.000000000   0.66346211
#>    16   91 0.647250785 0.499949400 0.509341582 0.468798987   0.57567863
#>    16   60 0.647250785 0.499949400 0.537330249 0.386097673   0.58633827
#>    16   19 0.647250785 0.499949400 0.561675206 0.365624646   0.59984758
#>    16   59 0.647250785 0.499949400 0.666038543 0.376036057   0.60789407
#>    16   64 0.647250785 0.499949400 0.728078340 0.468888663   0.67063906
#>    16   27 0.647250785 0.499949400 0.770443166 0.710702061   0.72949006
#>    16   35 0.647250785 0.499949400 0.588700301 0.566453094   0.70346821
#>    17   49 0.271721610 0.270166247 0.230178662 0.247645090   0.21451219
#>    17   45 0.271721610 0.270166247 0.322225162 0.237624211   0.27345408
#>    18   47 0.212554969 0.406119080 0.137534976 0.384636663   0.14020156
#>    19   56 0.561675206 0.365624646 0.453304542 0.267185767   0.50334253
#>    19   22 0.561675206 0.365624646 0.535002927 0.188181623   0.54258624
#>    19   97 0.561675206 0.365624646 0.678276151 0.177059113   0.61320195
#>    20   64 0.695003463 0.364346427 0.728078340 0.468888663   0.70086932
#>    20   59 0.695003463 0.364346427 0.666038543 0.376036057   0.70064846
#>    21   80 0.761369857 0.432357838 0.803430663 0.381820930   0.78302508
#>    21   94 0.761369857 0.432357838 0.781783347 0.446608092   0.78985736
#>    21   64 0.761369857 0.432357838 0.728078340 0.468888663   0.75638278
#>    21   72 0.761369857 0.432357838 0.751891435 0.397020281   0.71696732
#>    22   68 0.535002927 0.188181623 0.554109783 0.143858825   0.47468523
#>    23   41 0.339583644 0.596230240 0.248035250 0.593959346   0.29414267
#>    23   32 0.339583644 0.596230240 0.401579185 0.502992443   0.29616507
#>    23   71 0.339583644 0.596230240 0.466256784 0.654360851   0.42197118
#>    23   75 0.339583644 0.596230240 0.338789736 0.690845038   0.39433651
#>    23   66 0.339583644 0.596230240 0.257183710 0.611847019   0.30581663
#>    24   51 0.752007560 0.929325163 0.959902603 0.964856359   0.86217114
#>    25   82 0.858686902 0.652925260 0.866343397 0.597037594   0.76896319
#>    25   38 0.858686902 0.652925260 0.861718974 0.650491997   0.83581912
#>    25   52 0.858686902 0.652925260 0.867898308 0.710695995   0.88197010
#>    25   27 0.858686902 0.652925260 0.770443166 0.710702061   0.81916938
#>    26   39 0.522718789 0.852089056 0.420880882 0.844065074   0.47159156
#>    27   69 0.770443166 0.710702061 0.609199272 0.713156247   0.68928530
#>    28   34 0.719462453 0.147977926 0.718636695 0.160760272   0.75147477
#>    28   97 0.719462453 0.147977926 0.678276151 0.177059113   0.69187556
#>    28   73 0.719462453 0.147977926 0.590156032 0.099604268   0.66059815
#>    28   79 0.719462453 0.147977926 0.756920067 0.066998588   0.67230869
#>    28   65 0.719462453 0.147977926 0.799389037 0.125840015   0.75319971
#>    28   90 0.719462453 0.147977926 0.769346416 0.184345754   0.76116175
#>    29   53 0.166023769 0.676489733 0.156045936 0.691772411   0.18773288
#>    29   30 0.166023769 0.676489733 0.007691770 0.571691024   0.08149124
#>    29   41 0.166023769 0.676489733 0.248035250 0.593959346   0.13009218
#>    30   46 0.007691770 0.571691024 0.051185490 0.405536474  -0.60891096
#>    30   47 0.007691770 0.571691024 0.137534976 0.384636663   0.12292005
#>    30   42 0.007691770 0.571691024 0.003508729 0.868628470   0.01024020
#>    30   70 0.007691770 0.571691024 0.000000000 0.053160216 -14.09347368
#>    31   92 0.094788192 0.109384206 0.094388846 0.102442214   0.03059460
#>    31   62 0.094788192 0.109384206 0.215666137 0.009678677   0.18900486
#>    31   46 0.094788192 0.109384206 0.051185490 0.405536474   0.06060842
#>    31   70 0.094788192 0.109384206 0.000000000 0.053160216  -0.04666347
#>    32   91 0.401579185 0.502992443 0.509341582 0.468798987   0.43620647
#>    32   40 0.401579185 0.502992443 0.499491458 0.515347023   0.45403776
#>    33   73 0.412804968 0.000000000 0.590156032 0.099604268   0.62424852
#>    33   86 0.412804968 0.000000000 0.406186840 0.094917281   0.49928075
#>    33   61 0.412804968 0.000000000 0.323579525 0.127254642   0.33802494
#>    33   87 0.412804968 0.000000000 0.245446180 0.019742202   0.33251603
#>    33   62 0.412804968 0.000000000 0.215666137 0.009678677   0.30404360
#>    35   88 0.588700301 0.566453094 0.505748898 0.672198752   0.57720420
#>    35   71 0.588700301 0.566453094 0.466256784 0.654360851   0.51667895
#>    35   40 0.588700301 0.566453094 0.499491458 0.515347023   0.51457389
#>    37   89 0.982088838 0.481435555 0.969692975 0.219022641   1.07027311
#>    38   50 0.861718974 0.650491997 0.999658163 0.674645467   0.93638715
#>    39   43 0.420880882 0.844065074 0.386064477 0.837061138   0.40024745
#>    39   55 0.420880882 0.844065074 0.339691903 0.749124662   0.41897291
#>    39   71 0.420880882 0.844065074 0.466256784 0.654360851   0.43741492
#>    39   88 0.420880882 0.844065074 0.505748898 0.672198752   0.44683541
#>    44   78 1.000000000 0.861925934 0.917746909 0.796350520   0.91867313
#>    44   50 1.000000000 0.861925934 0.999658163 0.674645467   1.00739851
#>    57   64 0.872305019 0.521505965 0.728078340 0.468888663   0.77940503
#>    57   94 0.872305019 0.521505965 0.781783347 0.446608092   0.78628343
#>    57   63 0.872305019 0.521505965 0.877625580 0.432315711   0.83493288
#>    60  100 0.537330249 0.386097673 0.374982114 0.353656427   0.44963609
#>    62   74 0.215666137 0.009678677 0.107024285 0.018593816   0.16595819
#>    63   80 0.877625580 0.432315711 0.803430663 0.381820930   0.82768741
#>    63   93 0.877625580 0.432315711 0.907670486 0.416156342   0.86448326
#>    65   98 0.799389037 0.125840015 0.913519875 0.236911924   0.90943228
#>    68   73 0.554109783 0.143858825 0.590156032 0.099604268   0.49906428
#>    15   89 0.910001080 0.002755188 0.969692975 0.219022641   0.91436611
#>    44   54 1.000000000 0.861925934 0.967975720 0.946602668   0.91867313
#>    70   74 0.000000000 0.053160216 0.107024285 0.018593816   0.05946328
#>          my1          mx2         my2 bp1 bp2
#>   0.75231624   0.81917077  0.71126301   0   0
#>   0.78906366   0.84922607  0.82540926   0   0
#>   0.84206957   0.69931044  0.86680552   0   0
#>   0.82875493   0.63996942  0.83070944   0   0
#>   0.69464259   0.68928530  0.67671845   0   0
#>   0.18304327   0.28130724  0.18192901   0   0
#>   0.16997174   0.33802494  0.04247531   0   0
#>   0.16477629   0.47468523  0.13589989   0   0
#>   0.20452866   0.54258624  0.27776787   0   0
#>   0.27163841   0.46520530  0.34791957   0   0
#>   0.25695587   0.33841369  0.30027343   0   0
#>   0.72062368   0.41574614  0.71880055   0   0
#>   0.80789706   0.41897291  0.76351181   0   0
#>   0.84314651   0.40024745  0.85659564   0   0
#>   0.85303234   0.29983070  0.87765676   0   0
#>   0.79146006   0.11968350  0.82512624   0   0
#>   0.75157067   0.20097042  0.70202344   0   0
#>   0.71511649   0.29139548  0.65815487   0   0
#>   1.37374253  -1.50325060  5.47085207   0   1
#>   0.73638867   0.01024020  0.72022511   0   0
#>   0.73991380   0.18773288  0.70156182   0   0
#>   0.91127394   0.24644022  0.90814842   0   0
#>   0.93152359   0.43539650  0.92478905   0   0
#>   1.49952758  -1.15685396  7.12493288   0   1
#>   0.70202344   0.18773288  0.70156182   0   0
#>   0.64109123   0.20104343  0.62927609   0   0
#>   0.65815487   0.30581663  0.64325763   0   0
#>   0.21861460   0.69187556  0.15261362   0   0
#>   0.21551418   0.75147477  0.15646381   0   0
#>   0.21762703   0.84948944  0.18853280   0   0
#>   0.29935798   0.87742686  0.32376419   0   0
#>   0.32244220   0.78302508  0.40760941   0   0
#>   0.32233816   0.70086932  0.41999380   0   0
#>   0.27160704   0.64197935  0.27469156   0   0
#>   0.48961965   0.14020156  0.51705657   0   0
#>   0.32879882   0.21451219  0.32611979   0   0
#>   0.32146801   0.32947579  0.30433727   0   0
#>   0.35602035   0.44404950  0.40902501   0   0
#>   0.48239815   0.43620647  0.42521603   0   0
#>   0.49035263   0.29616507  0.50013054   0   0
#>   2.05217370   0.42302301  4.76050798   0   1
#>   0.91801124   0.40024745  0.85659564   0   0
#>   0.91493090   0.47159156  0.85072041   0   0
#>   0.94215519   0.58571817  0.88668309   0   0
#>   1.09816493   0.69775950  0.96896721   0   0
#>   0.82540926   0.86993224  0.76684483   0   0
#>   0.89583577   0.91867313  0.87956263   0   0
#>   0.90952552   0.86217114  0.91072025   0   0
#>   0.86680552   0.69775950  0.96896721   0   0
#>   0.58479689   0.93638715  0.63002442   0   0
#>   0.56515047   0.77940503  0.55217457   0   0
#>   0.50885301   0.91006408  0.47900462   0   0
#>   0.48059362   1.07027311  0.34577067   0   0
#>   0.56352580   6.22063936  0.10247764   0   0
#>   0.21960772   0.90943228  0.12693913   0   0
#>   0.34556541   1.07027311  0.34577067   0   0
#>   0.35025658   0.91535350  0.46645200   0   0
#>   0.32595993   0.86448326  0.37186964   0   0
#>   0.96896721   0.85423870  0.95713342   0   0
#>   0.88668309   0.47159156  0.85072041   0   0
#>   0.84739103   0.47696516  0.78252037   0   0
#>   0.83070944   0.58475387  0.74456755   0   0
#>   0.26567918   0.19548341  0.32400362   0   0
#>   0.25238632   0.06060842  0.25563786   0   0
#>   0.11199882   0.19791939  0.11128934   0   0
#>   0.11572391   0.33251603  0.03861270   0   0
#>   0.21580788   0.27345408  0.21739448   0   0
#>   0.11792189   0.90943228  0.12693913   0   0
#>   0.05147461   0.75319971  0.11443053   0   0
#>  -0.37020033   0.62424852 -0.16879331   0   0
#>   0.49596264   0.56475343  0.50484102   0   0
#>   0.44877026   0.44963609  0.40250604   0   0
#>   0.43572741   0.50334253  0.32097099   0   0
#>   0.43060116   0.62356080  0.27355877   0   0
#>   0.44011456   0.70064846  0.42006367   0   0
#>   0.59325911   0.74362097  0.59078343   0   0
#>   0.60846981   0.68853294  0.62728713   0   0
#>   0.30433727   0.33841369  0.30027343   0   0
#>   0.32611979   0.19548341  0.32400362   0   0
#>   0.51705657   0.13363598  0.52052250   0   0
#>   0.32097099   0.46520530  0.34791957   0   0
#>   0.26715329   0.60745582  0.19313519   0   0
#>   0.27355877   0.64197935  0.27469156   0   0
#>   0.41999380   0.71696732  0.42532776   0   0
#>   0.41329577   0.82768741  0.42593592   0   0
#>   0.46124812   0.78628343  0.53332053   0   0
#>   0.13589989   0.49906428  0.06221564   0   0
#>   0.19313519   0.61681645  0.15812721   0   0
#>   0.58166148   0.20104343  0.62927609   0   0
#>   0.58378135   0.44221211  0.57513268   0   0
#>   0.64400040   0.41574614  0.71880055   0   0
#>   0.91072025  46.68283734 21.17596882   0   1
#>   0.95713342   1.68581110  5.74989763   0   1
#>   0.61216500   0.74362097  0.59078343   0   0
#>   0.62132413   0.93638715  0.63002442   0   0
#>   0.67883254   0.92868023  0.67403822   0   0
#>   0.68884597   0.81917077  0.71126301   0   0
#>   0.78252037   0.46550803  0.75921489   0   0
#>   0.67671845   0.58475387  0.74456755   0   0
#>   0.62728713   0.57720420  0.64284320   0   0
#>   0.59078343   0.77940503  0.55217457   0   0
#>   0.71126301   0.86993224  0.76684483   0   0
#>   0.10831691   0.61681645  0.15812721   0   0
#>   0.07701377   0.62424852 -0.16879331   0   0
#>   0.14317665   0.84948944  0.18853280   0   0
#>   0.63219824   0.01024020  0.72022511   0   0
#>   0.55877095   0.13363598  0.52052250   0   0
#>   0.32151512  -0.04666347  0.23984414   0   0
#>   0.51308407   0.06060842  0.25563786   0   0
#>   0.52154211 -15.31974028  0.52681839   0   1
#>   0.10959453   0.05946328  0.05430291   0   0
#>   0.10048181   0.16595819  0.07035103   0   0
#>   0.11128934   0.30404360 -0.20275404   0   0
#>   0.42521603   0.44404950  0.40902501   0   0
#>   0.48141226   0.56475343  0.50484102   0   0
#>   0.57513268   0.51457389  0.59243255   0   0
#>   0.05371891   0.49906428  0.06221564   0   0
#>   0.64284320   0.54829627  0.71585881   0   0
#>   0.59536461   0.44683541  0.74999430   0   0
#>   0.74456755   0.54829627  0.71585881   0   0
#>   0.10247764   8.40598479 -0.04124713   0   1
#>   0.67403822   0.94251528  0.72460347   0   0
#>   0.74774099   0.41574614  0.71880055   0   0
#>   0.75921489   0.54829627  0.71585881   0   0
#>   0.76827188   0.94251528  0.72460347   0   0
#>   0.30027343   0.45858905  0.35770207   0   0
#>   0.40902501   0.44963609  0.40250604   0   0
#>   0.72460347   0.86993224  0.76684483   0   0
#>   0.34791957   0.45858905  0.35770207   0   0
#>   0.47452274   0.82768741  0.42593592   0   0
#>   0.47900462   0.91535350  0.46645200   0   0
#>   0.07035103   0.05946328  0.05430291   0   0
#>   0.75231624   0.77478029  0.78906366   0   0
#>   0.78906366   0.68242837  0.84206957   0   0
#>   0.84206957   0.65761316  0.82875493   0   0
#>   0.82875493   0.69225069  0.69464259   0   0
#>   0.69464259   0.77316085  0.75231624   0   0
#>   0.18304327   0.38793454  0.16997174   0   0
#>   0.16997174   0.45377838  0.16477629   0   0
#>   0.16477629   0.47176230  0.20452866   0   0
#>   0.20452866   0.42280407  0.27163841   0   0
#>   0.27163841   0.36931606  0.25695587   0   0
#>   0.25695587   0.37211028  0.18304327   0   0
#>   0.72062368   0.33480500  0.80789706   0   0
#>   0.80789706   0.31052346  0.84314651   0   0
#>   0.84314651   0.29966232  0.85303234   0   0
#>   0.85303234   0.21094283  0.79146006   0   0
#>   0.79146006   0.22541509  0.75157067   0   0
#>   0.75157067   0.28912256  0.71511649   0   0
#>   0.71511649   0.29797287  0.72062368   0   0
#>   1.37374253   0.02898069  0.73638867   0   0
#>   0.73638867   0.03362086  0.73991380   0   0
#>   0.73991380   0.11968350  0.82512624   0   0
#>   0.82512624   0.11905034  0.91127394   0   0
#>   0.91127394  -0.05322153  1.37374253   0   0
#>   0.90814842   0.29983070  0.87765676   0   0
#>   0.87765676   0.33609662  0.93152359   0   0
#>   0.93152359   0.17782399  1.49952758   0   0
#>   1.49952758   0.24644022  0.90814842   0   0
#>   0.70202344   0.20942169  0.64109123   0   0
#>   0.64109123   0.29139548  0.65815487   0   0
#>   0.21861460   0.72400999  0.21551418   0   0
#>   0.21551418   0.83888158  0.21762703   0   0
#>   0.21762703   0.84530128  0.29935798   0   0
#>   0.29935798   0.75790855  0.32244220   0   0
#>   0.32244220   0.75695826  0.32233816   0   0
#>   0.32233816   0.67651516  0.27160704   0   0
#>   0.27160704   0.71852881  0.21861460   0   0
#>   0.48961965   0.22066804  0.32879882   0   0
#>   0.32879882   0.31562492  0.32146801   0   0
#>   0.32146801   0.32481938  0.35602035   0   0
#>   0.35602035   0.29849881  0.48239815   0   0
#>   0.48239815   0.29037217  0.49035263   0   0
#>   0.49035263   0.28545851  0.48961965   0   0
#>   2.05217370   0.43539650  0.92478905   0   0
#>   0.92478905   0.44185938  0.91801124   0   0
#>   0.91801124   0.44763278  0.91493090   0   0
#>   0.91493090   0.56497596  0.94215519   0   0
#>   0.94215519   0.61480766  1.09816493   0   0
#>   1.09816493   0.53084772  2.05217370   0   0
#>   0.82540926   0.86999439  0.89583577   0   0
#>   0.89583577   0.86226672  0.90952552   0   0
#>   0.90952552   0.69931044  0.86680552   0   0
#>   0.58479689   0.94380519  0.56515047   0   0
#>   0.56515047   0.92657240  0.50885301   0   0
#>   0.50885301   0.97035411  0.48059362   0   0
#>   0.48059362   1.15049177  0.56352580   0   0
#>   0.56352580   0.96271591  0.58479689   0   0
#>   0.21960772   1.05581929  0.34556541   0   0
#>   0.34556541   1.02159419  0.35025658   0   0
#>   0.35025658   0.87960540  0.32595993   0   0
#>   0.32595993   0.87742686  0.32376419   0   0
#>   0.32376419   0.93894418  0.21960772   0   0
#>   0.88668309   0.61006464  0.84739103   0   0
#>   0.84739103   0.63996942  0.83070944   0   0
#>   0.26567918   0.08154551  0.25238632   0   0
#>   0.25238632   0.19834154  0.11199882   0   0
#>   0.11199882   0.22641007  0.11572391   0   0
#>   0.11572391   0.28130724  0.18192901   0   0
#>   0.18192901   0.27328135  0.21580788   0   0
#>   0.21580788   0.10923946  0.26567918   0   0
#>   0.11792189   0.84042613  0.05147461   0   0
#>   0.05147461   0.66346211 -0.37020033   0   0
#>  -0.37020033   0.67460717 -2.38141737   0   1
#>   0.49596264   0.58633827  0.44877026   0   0
#>   0.44877026   0.59984758  0.43572741   0   0
#>   0.43572741   0.60789407  0.43060116   0   0
#>   0.43060116   0.67063906  0.44011456   0   0
#>   0.44011456   0.72949006  0.59325911   0   0
#>   0.59325911   0.70346821  0.60846981   0   0
#>   0.60846981   0.57567863  0.49596264   0   0
#>   0.32611979   0.27345408  0.21739448   0   0
#>   0.21739448   0.32947579  0.30433727   0   0
#>   0.51705657   0.19548341  0.32400362   0   0
#>   0.32097099   0.54258624  0.27776787   0   0
#>   0.27776787   0.61320195  0.26715329   0   0
#>   0.26715329   0.62356080  0.27355877   0   0
#>   0.41999380   0.70064846  0.42006367   0   0
#>   0.42006367   0.64197935  0.27469156   0   0
#>   0.40760941   0.78985736  0.41329577   0   0
#>   0.41329577   0.75638278  0.46124812   0   0
#>   0.46124812   0.71696732  0.42532776   0   0
#>   0.42532776   0.78302508  0.40760941   0   0
#>   0.13589989   0.60745582  0.19313519   0   0
#>   0.58166148   0.29616507  0.50013054   0   0
#>   0.50013054   0.42197118  0.58378135   0   0
#>   0.58378135   0.39433651  0.64400040   0   0
#>   0.64400040   0.30581663  0.64325763   0   0
#>   0.64325763   0.29414267  0.58166148   0   0
#>   0.91072025   0.85423870  0.95713342   0   0
#>   0.61216500   0.83581912  0.62132413   0   0
#>   0.62132413   0.88197010  0.67883254   0   0
#>   0.67883254   0.81916938  0.68884597   0   0
#>   0.68884597   0.76896319  0.61216500   0   0
#>   0.85072041   0.47696516  0.78252037   0   0
#>   0.67671845   0.68853294  0.62728713   0   0
#>   0.15646381   0.69187556  0.15261362   0   0
#>   0.15261362   0.66059815  0.10831691   0   0
#>   0.10831691   0.67230869  0.07701377   0   0
#>   0.07701377   0.75319971  0.11443053   0   0
#>   0.11443053   0.76116175  0.14317665   0   0
#>   0.14317665   0.75147477  0.15646381   0   0
#>   0.70156182   0.08149124  0.63219824   0   0
#>   0.63219824   0.13009218  0.55877095   0   0
#>   0.55877095   0.20104343  0.62927609   0   0
#>   0.32151512   0.12292005  0.51308407   0   0
#>   0.51308407   0.13363598  0.52052250   0   0
#>   0.72022511 -14.09347368  0.52154211   0   0
#>   0.52154211  -0.60891096  0.32151512   0   0
#>   0.10959453   0.18900486  0.10048181   0   0
#>   0.10048181   0.19791939  0.11128934   0   0
#>   0.25563786  -0.04666347  0.23984414   0   0
#>   0.23984414   0.03059460  0.10959453   0   0
#>   0.42521603   0.45403776  0.48141226   0   0
#>   0.48141226   0.44221211  0.57513268   0   0
#>  -0.16879331   0.49928075  0.05371891   0   0
#>   0.05371891   0.33802494  0.04247531   0   0
#>   0.04247531   0.33251603  0.03861270   0   0
#>   0.03861270   0.30404360 -0.20275404   0   0
#>  -0.20275404   0.05560100 -5.26312393   0   1
#>   0.64284320   0.51667895  0.59536461   0   0
#>   0.59536461   0.51457389  0.59243255   0   0
#>   0.59243255   0.56475343  0.50484102   0   0
#>   0.34577067   6.22063936  0.10247764   0   0
#>   0.63002442   0.92868023  0.67403822   0   0
#>   0.85659564   0.41897291  0.76351181   0   0
#>   0.76351181   0.43741492  0.74774099   0   0
#>   0.74774099   0.44683541  0.74999430   0   0
#>   0.74999430   0.46550803  0.75921489   0   0
#>   0.87956263   1.00739851  0.76827188   0   0
#>   0.76827188   6.34696587  0.75852575   0   1
#>   0.55217457   0.78628343  0.53332053   0   0
#>   0.53332053   0.83493288  0.47452274   0   0
#>   0.47452274   0.91006408  0.47900462   0   0
#>   0.40250604   0.45858905  0.35770207   0   0
#>   0.07035103  -0.58431454 -9.07263758   0   1
#>   0.42593592   0.86448326  0.37186964   0   0
#>   0.37186964   0.91535350  0.46645200   0   0
#>   0.12693913   0.84948944  0.18853280   0   0
#>   0.06221564   0.61681645  0.15812721   0   0
#>   0.11792189   5.21095095 -1.06797679   0   1
#>   0.87956263  11.25051727  4.78700988   0   1
#>   0.05430291  -2.67326724 -8.40676127   0   1
#> 
#> $x
#>               [,1]        [,2]
#>   [1,] 0.732936104 0.763320582
#>   [2,] 0.417732104 0.241234791
#>   [3,] 0.289287958 0.770396726
#>   [4,] 0.020066181 0.870763954
#>   [5,] 0.263246131 0.950364352
#>   [6,] 0.241547212 0.686964806
#>   [7,] 0.768169570 0.248328717
#>   [8,] 0.280116055 0.378900504
#>   [9,] 0.491599796 0.976611775
#>  [10,] 0.775901911 0.838179689
#>  [11,] 0.978624040 0.488961469
#>  [12,] 0.936063778 0.250226896
#>  [13,] 0.644739798 0.927696709
#>  [14,] 0.219751917 0.213348346
#>  [15,] 0.910001080 0.002755188
#>  [16,] 0.647250785 0.499949400
#>  [17,] 0.271721610 0.270166247
#>  [18,] 0.212554969 0.406119080
#>  [19,] 0.561675206 0.365624646
#>  [20,] 0.695003463 0.364346427
#>  [21,] 0.761369857 0.432357838
#>  [22,] 0.535002927 0.188181623
#>  [23,] 0.339583644 0.596230240
#>  [24,] 0.752007560 0.929325163
#>  [25,] 0.858686902 0.652925260
#>  [26,] 0.522718789 0.852089056
#>  [27,] 0.770443166 0.710702061
#>  [28,] 0.719462453 0.147977926
#>  [29,] 0.166023769 0.676489733
#>  [30,] 0.007691770 0.571691024
#>  [31,] 0.094788192 0.109384206
#>  [32,] 0.401579185 0.502992443
#>  [33,] 0.412804968 0.000000000
#>  [34,] 0.718636695 0.160760272
#>  [35,] 0.588700301 0.566453094
#>  [36,] 0.623396784 0.735029583
#>  [37,] 0.982088838 0.481435555
#>  [38,] 0.861718974 0.650491997
#>  [39,] 0.420880882 0.844065074
#>  [40,] 0.499491458 0.515347023
#>  [41,] 0.248035250 0.593959346
#>  [42,] 0.003508729 0.868628470
#>  [43,] 0.386064477 0.837061138
#>  [44,] 1.000000000 0.861925934
#>  [45,] 0.322225162 0.237624211
#>  [46,] 0.051185490 0.405536474
#>  [47,] 0.137534976 0.384636663
#>  [48,] 0.356671577 0.394844640
#>  [49,] 0.230178662 0.247645090
#>  [50,] 0.999658163 0.674645467
#>  [51,] 0.959902603 0.964856359
#>  [52,] 0.867898308 0.710695995
#>  [53,] 0.156045936 0.691772411
#>  [54,] 0.967975720 0.946602668
#>  [55,] 0.339691903 0.749124662
#>  [56,] 0.453304542 0.267185767
#>  [57,] 0.872305019 0.521505965
#>  [58,] 0.164330003 0.725060761
#>  [59,] 0.666038543 0.376036057
#>  [60,] 0.537330249 0.386097673
#>  [61,] 0.323579525 0.127254642
#>  [62,] 0.215666137 0.009678677
#>  [63,] 0.877625580 0.432315711
#>  [64,] 0.728078340 0.468888663
#>  [65,] 0.799389037 0.125840015
#>  [66,] 0.257183710 0.611847019
#>  [67,] 0.380960267 0.871113149
#>  [68,] 0.554109783 0.143858825
#>  [69,] 0.609199272 0.713156247
#>  [70,] 0.000000000 0.053160216
#>  [71,] 0.466256784 0.654360851
#>  [72,] 0.751891435 0.397020281
#>  [73,] 0.590156032 0.099604268
#>  [74,] 0.107024285 0.018593816
#>  [75,] 0.338789736 0.690845038
#>  [76,] 0.814197680 0.759739423
#>  [77,] 0.218619241 0.872223266
#>  [78,] 0.917746909 0.796350520
#>  [79,] 0.756920067 0.066998588
#>  [80,] 0.803430663 0.381820930
#>  [81,] 0.388706036 0.985323385
#>  [82,] 0.866343397 0.597037594
#>  [83,] 0.514267130 0.878910172
#>  [84,] 0.966134396 0.469611516
#>  [85,] 0.757352368 1.000000000
#>  [86,] 0.406186840 0.094917281
#>  [87,] 0.245446180 0.019742202
#>  [88,] 0.505748898 0.672198752
#>  [89,] 0.969692975 0.219022641
#>  [90,] 0.769346416 0.184345754
#>  [91,] 0.509341582 0.468798987
#>  [92,] 0.094388846 0.102442214
#>  [93,] 0.907670486 0.416156342
#>  [94,] 0.781783347 0.446608092
#>  [95,] 0.559950954 0.775697329
#>  [96,] 0.402920826 0.295191783
#>  [97,] 0.678276151 0.177059113
#>  [98,] 0.913519875 0.236911924
#>  [99,] 0.220414505 0.945394707
#> [100,] 0.374982114 0.353656427
#> 
#> $tri.obj
#> triangulation nodes with neigbours:
#> node: (x,y): neighbours
#> 1: (0.7329361,0.7633206) [5]: 10 13 27 36 76 
#> 2: (0.4177321,0.2412348) [6]: 22 45 56 61 86 96 
#> 3: (0.289288,0.7703967) [7]: 6 43 55 58 67 75 77 
#> 4: (0.02006618,0.870764) [5]: 42 53 58 77 99 
#> 5: (0.2632461,0.9503644) [4]: 67 77 81 99 
#> 6: (0.2415472,0.6869648) [5]: 3 29 58 66 75 
#> 7: (0.7681696,0.2483287) [7]: 20 34 72 80 90 97 98 
#> 8: (0.2801161,0.3789005) [6]: 17 18 32 41 48 100 
#> 9: (0.4915998,0.9766118) [6]: 13 39 67 81 83 85 
#> 10: (0.7759019,0.8381797) [6]: 1 13 24 54 76 78 
#> 11: (0.978624,0.4889615) [5]: 37 50 57 82 84 
#> 12: (0.9360638,0.2502269) [5]: 80 84 89 93 98 
#> 13: (0.6447398,0.9276967) [9]: 1 9 10 24 26 36 83 85 95 
#> 14: (0.2197519,0.2133483) [6]: 31 45 47 49 61 87 
#> 15: (0.9100011,0.002755188) [4]: 33 65 79 89 
#> 16: (0.6472508,0.4999494) [7]: 19 27 35 59 60 64 91 
#> 17: (0.2717216,0.2701662) [5]: 8 18 45 49 100 
#> 18: (0.212555,0.4061191) [5]: 8 17 41 47 49 
#> 19: (0.5616752,0.3656246) [6]: 16 22 56 59 60 97 
#> 20: (0.6950035,0.3643464) [5]: 7 59 64 72 97 
#> 21: (0.7613699,0.4323578) [4]: 64 72 80 94 
#> 22: (0.5350029,0.1881816) [6]: 2 19 56 68 86 97 
#> 23: (0.3395836,0.5962302) [5]: 32 41 66 71 75 
#> 24: (0.7520076,0.9293252) [5]: 10 13 51 54 85 
#> 25: (0.8586869,0.6529253) [4]: 27 38 52 82 
#> 26: (0.5227188,0.8520891) [4]: 13 39 83 95 
#> 27: (0.7704432,0.7107021) [10]: 1 16 25 35 36 52 64 69 76 82 
#> 28: (0.7194625,0.1479779) [6]: 34 65 73 79 90 97 
#> 29: (0.1660238,0.6764897) [6]: 6 30 41 53 58 66 
#> 30: (0.00769177,0.571691) [7]: 29 41 42 46 47 53 70 
#> 31: (0.09478819,0.1093842) [7]: 14 46 47 62 70 87 92 
#> 32: (0.4015792,0.5029924) [7]: 8 23 40 41 48 71 91 
#> 33: (0.412805,0) [7]: 15 61 62 73 79 86 87 
#> 34: (0.7186367,0.1607603) [4]: 7 28 90 97 
#> 35: (0.5887003,0.5664531) [7]: 16 27 40 69 71 88 91 
#> 36: (0.6233968,0.7350296) [5]: 1 13 27 69 95 
#> 37: (0.9820888,0.4814356) [4]: 11 50 84 89 
#> 38: (0.861719,0.650492) [4]: 25 50 52 82 
#> 39: (0.4208809,0.8440651) [9]: 9 26 43 55 67 71 83 88 95 
#> 40: (0.4994915,0.515347) [4]: 32 35 71 91 
#> 41: (0.2480352,0.5939593) [8]: 8 18 23 29 30 32 47 66 
#> 42: (0.003508729,0.8686285) [5]: 4 30 53 70 99 
#> 43: (0.3860645,0.8370611) [4]: 3 39 55 67 
#> 44: (1,0.8619259) [3]: 50 54 78 
#> 45: (0.3222252,0.2376242) [7]: 2 14 17 49 61 96 100 
#> 46: (0.05118549,0.4055365) [4]: 30 31 47 70 
#> 47: (0.137535,0.3846367) [7]: 14 18 30 31 41 46 49 
#> 48: (0.3566716,0.3948446) [4]: 8 32 91 100 
#> 49: (0.2301787,0.2476451) [5]: 14 17 18 45 47 
#> 50: (0.9996582,0.6746455) [8]: 11 37 38 44 52 78 82 89 
#> 51: (0.9599026,0.9648564) [3]: 24 54 85 
#> 52: (0.8678983,0.710696) [6]: 25 27 38 50 76 78 
#> 53: (0.1560459,0.6917724) [5]: 4 29 30 42 58 
#> 54: (0.9679757,0.9466027) [5]: 10 24 44 51 78 
#> 55: (0.3396919,0.7491247) [5]: 3 39 43 71 75 
#> 56: (0.4533045,0.2671858) [5]: 2 19 22 60 96 
#> 57: (0.872305,0.521506) [6]: 11 63 64 82 84 94 
#> 58: (0.16433,0.7250608) [6]: 3 4 6 29 53 77 
#> 59: (0.6660385,0.3760361) [5]: 16 19 20 64 97 
#> 60: (0.5373302,0.3860977) [6]: 16 19 56 91 96 100 
#> 61: (0.3235795,0.1272546) [6]: 2 14 33 45 86 87 
#> 62: (0.2156661,0.009678677) [5]: 31 33 74 87 92 
#> 63: (0.8776256,0.4323157) [5]: 57 80 84 93 94 
#> 64: (0.7280783,0.4688887) [9]: 16 20 21 27 57 59 72 82 94 
#> 65: (0.799389,0.12584) [6]: 15 28 79 89 90 98 
#> 66: (0.2571837,0.611847) [5]: 6 23 29 41 75 
#> 67: (0.3809603,0.8711131) [7]: 3 5 9 39 43 77 81 
#> 68: (0.5541098,0.1438588) [4]: 22 73 86 97 
#> 69: (0.6091993,0.7131562) [5]: 27 35 36 88 95 
#> 70: (0,0.05316022) [6]: 30 31 42 46 74 92 
#> 71: (0.4662568,0.6543609) [8]: 23 32 35 39 40 55 75 88 
#> 72: (0.7518914,0.3970203) [5]: 7 20 21 64 80 
#> 73: (0.590156,0.09960427) [6]: 28 33 68 79 86 97 
#> 74: (0.1070243,0.01859382) [3]: 62 70 92 
#> 75: (0.3387897,0.690845) [6]: 3 6 23 55 66 71 
#> 76: (0.8141977,0.7597394) [5]: 1 10 27 52 78 
#> 77: (0.2186192,0.8722233) [6]: 3 4 5 58 67 99 
#> 78: (0.9177469,0.7963505) [6]: 10 44 50 52 54 76 
#> 79: (0.7569201,0.06699859) [5]: 15 28 33 65 73 
#> 80: (0.8034307,0.3818209) [8]: 7 12 21 63 72 93 94 98 
#> 81: (0.388706,0.9853234) [5]: 5 9 67 85 99 
#> 82: (0.8663434,0.5970376) [7]: 11 25 27 38 50 57 64 
#> 83: (0.5142671,0.8789102) [4]: 9 13 26 39 
#> 84: (0.9661344,0.4696115) [7]: 11 12 37 57 63 89 93 
#> 85: (0.7573524,1) [5]: 9 13 24 51 81 
#> 86: (0.4061868,0.09491728) [6]: 2 22 33 61 68 73 
#> 87: (0.2454462,0.0197422) [5]: 14 31 33 61 62 
#> 88: (0.5057489,0.6721988) [5]: 35 39 69 71 95 
#> 89: (0.969693,0.2190226) [7]: 12 15 37 50 65 84 98 
#> 90: (0.7693464,0.1843458) [5]: 7 28 34 65 98 
#> 91: (0.5093416,0.468799) [7]: 16 32 35 40 48 60 100 
#> 92: (0.09438885,0.1024422) [4]: 31 62 70 74 
#> 93: (0.9076705,0.4161563) [4]: 12 63 80 84 
#> 94: (0.7817833,0.4466081) [5]: 21 57 63 64 80 
#> 95: (0.559951,0.7756973) [6]: 13 26 36 39 69 88 
#> 96: (0.4029208,0.2951918) [5]: 2 45 56 60 100 
#> 97: (0.6782762,0.1770591) [9]: 7 19 20 22 28 34 59 68 73 
#> 98: (0.9135199,0.2369119) [6]: 7 12 65 80 89 90 
#> 99: (0.2204145,0.9453947) [5]: 4 5 42 77 81 
#> 100: (0.3749821,0.3536564) [7]: 8 17 45 48 60 91 96 
#> number of nodes: 100 
#> number of arcs: 283 
#> number of boundary nodes: 14 
#> boundary nodes:  15 33 42 44 50 51 54 62 70 74 81 85 89 99 
#> number of triangles: 184 
#> number of constraints: 0 
#> 
#> attr(,"class")
#> [1] "delvor"
#> 
#> $weights
#>                                                                         
#> 0.065720015 0.087289411 0.158797975 0.193845675 0.149045184 0.110377879 
#>                                                                         
#> 0.088711179 0.159034027 0.113649818 0.057644216 0.099125252 0.058286607 
#>                                                                         
#> 0.099414482 0.034432432 0.162344821 0.156857019 0.086103422 0.097319909 
#>                                                                         
#> 0.230089485 0.233550134 0.034303644 0.073193461 0.114472595 0.172963403 
#>                                                                         
#> 0.048600552 0.111753368 0.113579182 0.043527299 0.055926297 0.153457449 
#>                                                                         
#> 0.181984216 0.053733716 0.065603522 0.188032819 0.191161753 0.148269566 
#>                                                                         
#> 0.132790594 0.045074878 0.117100985 0.178467693 0.368938370 0.048220887 
#>                                                                         
#> 0.099675333 0.139295526 0.133825845 0.109830848 0.158425508 0.216658160 
#>                                                                         
#> 0.107280122 0.154258902 0.075766536 0.107223998 0.019858300 0.194007091 
#>                                                                         
#> 0.058952892 0.250614141 0.079217955 0.109749083 0.070876650 0.028121216 
#>                                                                         
#> 0.084981940 0.075360725 0.165376975 0.278551975 0.175309778 0.132905020 
#>                                                                         
#> 0.092590367 0.194129939 0.072566707 0.350576685 0.125833743 0.087309066 
#>                                                                         
#> 0.031809145 0.104881379 0.111671585 0.245496431 0.232030674 0.127462822 
#>                                                                         
#> 0.159450933 0.236698745 0.145603437 0.143704303 0.199352911 0.075710817 
#>                                                                         
#> 0.068307998 0.058143372 0.155809088 0.128528386 0.020091370 0.164607371 
#>                                                                         
#> 0.132585610 0.019959270 0.205576441 0.148716057 0.053654063 0.060520290 
#>                                                                         
#> 0.097455142 0.154966555 0.026077043 0.148128400 0.188518897 0.072725616 
#>                                                                         
#> 0.117321815 0.169921669 0.065768386 0.190862511 0.241372878 0.356074404 
#>                                                                         
#> 0.088842759 0.815475802 0.106479903 0.152686787 0.031434464 0.169638957 
#>                                                                         
#> 0.047578825 0.142931409 0.184028888 0.111263185 0.043333795 0.079603918 
#>                                                                         
#> 0.456607131 0.136602709 0.158110243 0.116832437 0.146702329 0.064797271 
#>                                                                         
#> 0.176947098 0.099103888 0.162264518 0.096902042 0.096045758 0.084795095 
#>                                                                         
#> 0.081340448 0.086313073 0.186542487 0.113133741 0.064618019 0.147838366 
#>                                                                         
#> 0.146772296 0.128713188 0.044032391 0.055952935 0.095575165 0.054708851 
#>                                                                         
#> 0.117515269 0.136189611 0.123946407 0.132927951 0.096125252 0.093695769 
#>                                                                         
#> 0.016694596 0.224785378 0.205040168 0.198558422 0.213797100 0.089986602 
#>                                                                         
#> 0.141906205 0.130239479 0.043118970 0.076246427 0.076727974 0.100606850 
#>                                                                         
#> 0.063993785 0.145797992 0.138070690 0.149579941 0.137161905 0.114717842 
#>                                                                         
#> 0.072837842 0.109057807 0.098167370 0.078198232 0.173643605 0.217438459 
#>                                                                         
#> 0.103261890 0.152875982 0.150232463 0.100296617 0.160762383 0.266779756 
#>                                                                         
#> 0.147884018 0.220562667 0.094225461 0.155844125 0.111188482 0.023030673 
#>                                                                         
#> 0.008285180 0.186871564 0.045876229 0.221435891 0.168341202 0.186838254 
#>                                                                         
#> 0.026182361 0.143546662 0.174048728 0.189998192 0.162556067 0.195303698 
#>                                                                         
#> 0.134878828 0.105309466 0.035846669 0.165483832 0.166015093 0.497203747 
#>                                                                         
#> 0.141383509 0.158255300 0.159268073 0.125329551 0.086590202 0.244116872 
#>                                                                         
#> 0.088605307 0.047254830 0.060079888 0.078035207 0.146404965 0.179436443 
#>                                                                         
#> 0.221704174 0.109649563 0.031234821 0.065750212 0.024895388 0.049424956 
#>                                                                         
#> 0.036586656 0.048265747 0.091576555 0.111967557 0.139374503 0.094618129 
#>                                                                         
#> 0.083866757 0.210909495 0.056409690 0.003887703 0.058500494 0.105475664 
#>                                                                         
#> 0.102153529 0.161262570 0.012808991 0.050418518 0.138058543 0.089222900 
#>                                                                         
#> 0.082935794 0.061733530 0.018251505 0.189873092 0.116349249 0.171752841 
#>                                                                         
#> 0.227702859 0.296966908 0.518587854 0.006953469 0.156692917 0.299344886 
#>                                                                         
#> 0.110208613 0.113057182 0.098688646 0.203407007 0.095147727 0.155418543 
#>                                                                         
#> 0.168519193 0.197376279 0.134398956 0.150732175 0.102810739 0.262705528 
#>                                                                         
#> 0.140037888 0.035513902 0.124921303 0.195055542 0.191678410 0.105193659 
#>                                                                         
#> 0.187280779 0.153524966 0.117489848 0.089348809 0.165557698 0.109007026 
#>                                                                         
#> 0.089747472 0.034114830 0.159257079 0.057077122 0.224354036 0.090530126 
#>             
#> 0.112467922 
#> 
#> $alpha
#> [1] 0.2908922
#> 
#> attr(,"class")
#> [1] "scree"